Enum unicode_types::alphabetic_presentation_forms::AlphabeticPresentationForms

pub enum AlphabeticPresentationForms {
    LatinSmallLigatureFf,
    LatinSmallLigatureFi,
    LatinSmallLigatureFl,
    LatinSmallLigatureFfi,
    LatinSmallLigatureFfl,
    LatinSmallLigatureLongST,
    LatinSmallLigatureSt,
    ArmenianSmallLigatureMenNow,
    ArmenianSmallLigatureMenEch,
    ArmenianSmallLigatureMenIni,
    ArmenianSmallLigatureVewNow,
    ArmenianSmallLigatureMenXeh,
    HebrewLetterYodWithHiriq,
    HebrewPointJudeoDashSpanishVarika,
    HebrewLigatureYiddishYodYodPatah,
    HebrewLetterAlternativeAyin,
    HebrewLetterWideAlef,
    HebrewLetterWideDalet,
    HebrewLetterWideHe,
    HebrewLetterWideKaf,
    HebrewLetterWideLamed,
    HebrewLetterWideFinalMem,
    HebrewLetterWideResh,
    HebrewLetterWideTav,
    HebrewLetterAlternativePlusSign,
    HebrewLetterShinWithShinDot,
    HebrewLetterShinWithSinDot,
    HebrewLetterShinWithDageshAndShinDot,
    HebrewLetterShinWithDageshAndSinDot,
    HebrewLetterAlefWithPatah,
    HebrewLetterAlefWithQamats,
    HebrewLetterAlefWithMapiq,
    HebrewLetterBetWithDagesh,
    HebrewLetterGimelWithDagesh,
    HebrewLetterDaletWithDagesh,
    HebrewLetterHeWithMapiq,
    HebrewLetterVavWithDagesh,
    HebrewLetterZayinWithDagesh,
    HebrewLetterTetWithDagesh,
    HebrewLetterYodWithDagesh,
    HebrewLetterFinalKafWithDagesh,
    HebrewLetterKafWithDagesh,
    HebrewLetterLamedWithDagesh,
    HebrewLetterMemWithDagesh,
    HebrewLetterNunWithDagesh,
    HebrewLetterSamekhWithDagesh,
    HebrewLetterFinalPeWithDagesh,
    HebrewLetterPeWithDagesh,
    HebrewLetterTsadiWithDagesh,
    HebrewLetterQofWithDagesh,
    HebrewLetterReshWithDagesh,
    HebrewLetterShinWithDagesh,
    HebrewLetterTavWithDagesh,
    HebrewLetterVavWithHolam,
    HebrewLetterBetWithRafe,
    HebrewLetterKafWithRafe,
    HebrewLetterPeWithRafe,
}

An enum to represent all characters in the AlphabeticPresentationForms block.

Variants
LatinSmallLigatureFf

\u{fb00}: 'ﬀ'

LatinSmallLigatureFi

\u{fb01}: 'ﬁ'

LatinSmallLigatureFl

\u{fb02}: 'ﬂ'

LatinSmallLigatureFfi

\u{fb03}: 'ﬃ'

LatinSmallLigatureFfl

\u{fb04}: 'ﬄ'

LatinSmallLigatureLongST

\u{fb05}: 'ﬅ'

LatinSmallLigatureSt

\u{fb06}: 'ﬆ'

ArmenianSmallLigatureMenNow

\u{fb13}: 'ﬓ'

ArmenianSmallLigatureMenEch

\u{fb14}: 'ﬔ'

ArmenianSmallLigatureMenIni

\u{fb15}: 'ﬕ'

ArmenianSmallLigatureVewNow

\u{fb16}: 'ﬖ'

ArmenianSmallLigatureMenXeh

\u{fb17}: 'ﬗ'

HebrewLetterYodWithHiriq

\u{fb1d}: 'יִ'

HebrewPointJudeoDashSpanishVarika

\u{fb1e}: 'ﬞ'

HebrewLigatureYiddishYodYodPatah

\u{fb1f}: 'ײַ'

HebrewLetterAlternativeAyin

\u{fb20}: 'ﬠ'

HebrewLetterWideAlef

\u{fb21}: 'ﬡ'

HebrewLetterWideDalet

\u{fb22}: 'ﬢ'

HebrewLetterWideHe

\u{fb23}: 'ﬣ'

HebrewLetterWideKaf

\u{fb24}: 'ﬤ'

HebrewLetterWideLamed

\u{fb25}: 'ﬥ'

HebrewLetterWideFinalMem

\u{fb26}: 'ﬦ'

HebrewLetterWideResh

\u{fb27}: 'ﬧ'

HebrewLetterWideTav

\u{fb28}: 'ﬨ'

HebrewLetterAlternativePlusSign

\u{fb29}: '﬩'

HebrewLetterShinWithShinDot

\u{fb2a}: 'שׁ'

HebrewLetterShinWithSinDot

\u{fb2b}: 'שׂ'

HebrewLetterShinWithDageshAndShinDot

\u{fb2c}: 'שּׁ'

HebrewLetterShinWithDageshAndSinDot

\u{fb2d}: 'שּׂ'

HebrewLetterAlefWithPatah

\u{fb2e}: 'אַ'

HebrewLetterAlefWithQamats

\u{fb2f}: 'אָ'

HebrewLetterAlefWithMapiq

\u{fb30}: 'אּ'

HebrewLetterBetWithDagesh

\u{fb31}: 'בּ'

HebrewLetterGimelWithDagesh

\u{fb32}: 'גּ'

HebrewLetterDaletWithDagesh

\u{fb33}: 'דּ'

HebrewLetterHeWithMapiq

\u{fb34}: 'הּ'

HebrewLetterVavWithDagesh

\u{fb35}: 'וּ'

HebrewLetterZayinWithDagesh

\u{fb36}: 'זּ'

HebrewLetterTetWithDagesh

\u{fb38}: 'טּ'

HebrewLetterYodWithDagesh

\u{fb39}: 'יּ'

HebrewLetterFinalKafWithDagesh

\u{fb3a}: 'ךּ'

HebrewLetterKafWithDagesh

\u{fb3b}: 'כּ'

HebrewLetterLamedWithDagesh

\u{fb3c}: 'לּ'

HebrewLetterMemWithDagesh

\u{fb3e}: 'מּ'

HebrewLetterNunWithDagesh

\u{fb40}: 'נּ'

HebrewLetterSamekhWithDagesh

\u{fb41}: 'סּ'

HebrewLetterFinalPeWithDagesh

\u{fb43}: 'ףּ'

HebrewLetterPeWithDagesh

\u{fb44}: 'פּ'

HebrewLetterTsadiWithDagesh

\u{fb46}: 'צּ'

HebrewLetterQofWithDagesh

\u{fb47}: 'קּ'

HebrewLetterReshWithDagesh

\u{fb48}: 'רּ'

HebrewLetterShinWithDagesh

\u{fb49}: 'שּ'

HebrewLetterTavWithDagesh

\u{fb4a}: 'תּ'

HebrewLetterVavWithHolam

\u{fb4b}: 'וֹ'

HebrewLetterBetWithRafe

\u{fb4c}: 'בֿ'

HebrewLetterKafWithRafe

\u{fb4d}: 'כֿ'

HebrewLetterPeWithRafe

\u{fb4e}: 'פֿ'

Methods

impl AlphabeticPresentationForms

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case