Enum unicode_types::anatolian_hieroglyphs::AnatolianHieroglyphs

pub enum AnatolianHieroglyphs {
    AnatolianHieroglyphA001,
    AnatolianHieroglyphA002,
    AnatolianHieroglyphA003,
    AnatolianHieroglyphA004,
    AnatolianHieroglyphA005,
    AnatolianHieroglyphA006,
    AnatolianHieroglyphA007,
    AnatolianHieroglyphA008,
    AnatolianHieroglyphA009,
    AnatolianHieroglyphA010,
    AnatolianHieroglyphA010a,
    AnatolianHieroglyphA011,
    AnatolianHieroglyphA012,
    AnatolianHieroglyphA013,
    AnatolianHieroglyphA014,
    AnatolianHieroglyphA015,
    AnatolianHieroglyphA016,
    AnatolianHieroglyphA017,
    AnatolianHieroglyphA018,
    AnatolianHieroglyphA019,
    AnatolianHieroglyphA020,
    AnatolianHieroglyphA021,
    AnatolianHieroglyphA022,
    AnatolianHieroglyphA023,
    AnatolianHieroglyphA024,
    AnatolianHieroglyphA025,
    AnatolianHieroglyphA026,
    AnatolianHieroglyphA026a,
    AnatolianHieroglyphA027,
    AnatolianHieroglyphA028,
    AnatolianHieroglyphA029,
    AnatolianHieroglyphA030,
    AnatolianHieroglyphA031,
    AnatolianHieroglyphA032,
    AnatolianHieroglyphA033,
    AnatolianHieroglyphA034,
    AnatolianHieroglyphA035,
    AnatolianHieroglyphA036,
    AnatolianHieroglyphA037,
    AnatolianHieroglyphA038,
    AnatolianHieroglyphA039,
    AnatolianHieroglyphA039a,
    AnatolianHieroglyphA040,
    AnatolianHieroglyphA041,
    AnatolianHieroglyphA041a,
    AnatolianHieroglyphA042,
    AnatolianHieroglyphA043,
    AnatolianHieroglyphA044,
    AnatolianHieroglyphA045,
    AnatolianHieroglyphA045a,
    AnatolianHieroglyphA046,
    AnatolianHieroglyphA046a,
    AnatolianHieroglyphA046b,
    AnatolianHieroglyphA047,
    AnatolianHieroglyphA048,
    AnatolianHieroglyphA049,
    AnatolianHieroglyphA050,
    AnatolianHieroglyphA051,
    AnatolianHieroglyphA052,
    AnatolianHieroglyphA053,
    AnatolianHieroglyphA054,
    AnatolianHieroglyphA055,
    AnatolianHieroglyphA056,
    AnatolianHieroglyphA057,
    AnatolianHieroglyphA058,
    AnatolianHieroglyphA059,
    AnatolianHieroglyphA060,
    AnatolianHieroglyphA061,
    AnatolianHieroglyphA062,
    AnatolianHieroglyphA063,
    AnatolianHieroglyphA064,
    AnatolianHieroglyphA065,
    AnatolianHieroglyphA066,
    AnatolianHieroglyphA066a,
    AnatolianHieroglyphA066b,
    AnatolianHieroglyphA066c,
    AnatolianHieroglyphA067,
    AnatolianHieroglyphA068,
    AnatolianHieroglyphA069,
    AnatolianHieroglyphA070,
    AnatolianHieroglyphA071,
    AnatolianHieroglyphA072,
    AnatolianHieroglyphA073,
    AnatolianHieroglyphA074,
    AnatolianHieroglyphA075,
    AnatolianHieroglyphA076,
    AnatolianHieroglyphA077,
    AnatolianHieroglyphA078,
    AnatolianHieroglyphA079,
    AnatolianHieroglyphA080,
    AnatolianHieroglyphA081,
    AnatolianHieroglyphA082,
    AnatolianHieroglyphA083,
    AnatolianHieroglyphA084,
    AnatolianHieroglyphA085,
    AnatolianHieroglyphA086,
    AnatolianHieroglyphA087,
    AnatolianHieroglyphA088,
    AnatolianHieroglyphA089,
    AnatolianHieroglyphA090,
    AnatolianHieroglyphA091,
    AnatolianHieroglyphA092,
    AnatolianHieroglyphA093,
    AnatolianHieroglyphA094,
    AnatolianHieroglyphA095,
    AnatolianHieroglyphA096,
    AnatolianHieroglyphA097,
    AnatolianHieroglyphA097a,
    AnatolianHieroglyphA098,
    AnatolianHieroglyphA098a,
    AnatolianHieroglyphA099,
    AnatolianHieroglyphA100,
    AnatolianHieroglyphA100a,
    AnatolianHieroglyphA101,
    AnatolianHieroglyphA101a,
    AnatolianHieroglyphA102,
    AnatolianHieroglyphA102a,
    AnatolianHieroglyphA103,
    AnatolianHieroglyphA104,
    AnatolianHieroglyphA104a,
    AnatolianHieroglyphA104b,
    AnatolianHieroglyphA104c,
    AnatolianHieroglyphA105,
    AnatolianHieroglyphA105a,
    AnatolianHieroglyphA105b,
    AnatolianHieroglyphA106,
    AnatolianHieroglyphA107,
    AnatolianHieroglyphA107a,
    AnatolianHieroglyphA107b,
    AnatolianHieroglyphA107c,
    AnatolianHieroglyphA108,
    AnatolianHieroglyphA109,
    AnatolianHieroglyphA110,
    AnatolianHieroglyphA110a,
    AnatolianHieroglyphA110b,
    AnatolianHieroglyphA111,
    AnatolianHieroglyphA112,
    AnatolianHieroglyphA113,
    AnatolianHieroglyphA114,
    AnatolianHieroglyphA115,
    AnatolianHieroglyphA115a,
    AnatolianHieroglyphA116,
    AnatolianHieroglyphA117,
    AnatolianHieroglyphA118,
    AnatolianHieroglyphA119,
    AnatolianHieroglyphA120,
    AnatolianHieroglyphA121,
    AnatolianHieroglyphA122,
    AnatolianHieroglyphA123,
    AnatolianHieroglyphA124,
    AnatolianHieroglyphA125,
    AnatolianHieroglyphA125a,
    AnatolianHieroglyphA126,
    AnatolianHieroglyphA127,
    AnatolianHieroglyphA128,
    AnatolianHieroglyphA129,
    AnatolianHieroglyphA130,
    AnatolianHieroglyphA131,
    AnatolianHieroglyphA132,
    AnatolianHieroglyphA133,
    AnatolianHieroglyphA134,
    AnatolianHieroglyphA135,
    AnatolianHieroglyphA135a,
    AnatolianHieroglyphA136,
    AnatolianHieroglyphA137,
    AnatolianHieroglyphA138,
    AnatolianHieroglyphA139,
    AnatolianHieroglyphA140,
    AnatolianHieroglyphA141,
    AnatolianHieroglyphA142,
    AnatolianHieroglyphA143,
    AnatolianHieroglyphA144,
    AnatolianHieroglyphA145,
    AnatolianHieroglyphA146,
    AnatolianHieroglyphA147,
    AnatolianHieroglyphA148,
    AnatolianHieroglyphA149,
    AnatolianHieroglyphA150,
    AnatolianHieroglyphA151,
    AnatolianHieroglyphA152,
    AnatolianHieroglyphA153,
    AnatolianHieroglyphA154,
    AnatolianHieroglyphA155,
    AnatolianHieroglyphA156,
    AnatolianHieroglyphA157,
    AnatolianHieroglyphA158,
    AnatolianHieroglyphA159,
    AnatolianHieroglyphA160,
    AnatolianHieroglyphA161,
    AnatolianHieroglyphA162,
    AnatolianHieroglyphA163,
    AnatolianHieroglyphA164,
    AnatolianHieroglyphA165,
    AnatolianHieroglyphA166,
    AnatolianHieroglyphA167,
    AnatolianHieroglyphA168,
    AnatolianHieroglyphA169,
    AnatolianHieroglyphA170,
    AnatolianHieroglyphA171,
    AnatolianHieroglyphA172,
    AnatolianHieroglyphA173,
    AnatolianHieroglyphA174,
    AnatolianHieroglyphA175,
    AnatolianHieroglyphA176,
    AnatolianHieroglyphA177,
    AnatolianHieroglyphA178,
    AnatolianHieroglyphA179,
    AnatolianHieroglyphA180,
    AnatolianHieroglyphA181,
    AnatolianHieroglyphA182,
    AnatolianHieroglyphA183,
    AnatolianHieroglyphA184,
    AnatolianHieroglyphA185,
    AnatolianHieroglyphA186,
    AnatolianHieroglyphA187,
    AnatolianHieroglyphA188,
    AnatolianHieroglyphA189,
    AnatolianHieroglyphA190,
    AnatolianHieroglyphA191,
    AnatolianHieroglyphA192,
    AnatolianHieroglyphA193,
    AnatolianHieroglyphA194,
    AnatolianHieroglyphA195,
    AnatolianHieroglyphA196,
    AnatolianHieroglyphA197,
    AnatolianHieroglyphA198,
    AnatolianHieroglyphA199,
    AnatolianHieroglyphA200,
    AnatolianHieroglyphA201,
    AnatolianHieroglyphA202,
    AnatolianHieroglyphA202a,
    AnatolianHieroglyphA202b,
    AnatolianHieroglyphA203,
    AnatolianHieroglyphA204,
    AnatolianHieroglyphA205,
    AnatolianHieroglyphA206,
    AnatolianHieroglyphA207,
    AnatolianHieroglyphA207a,
    AnatolianHieroglyphA208,
    AnatolianHieroglyphA209,
    AnatolianHieroglyphA209a,
    AnatolianHieroglyphA210,
    AnatolianHieroglyphA211,
    AnatolianHieroglyphA212,
    AnatolianHieroglyphA213,
    AnatolianHieroglyphA214,
    AnatolianHieroglyphA215,
    AnatolianHieroglyphA215a,
    AnatolianHieroglyphA216,
    AnatolianHieroglyphA216a,
    AnatolianHieroglyphA217,
    AnatolianHieroglyphA218,
    AnatolianHieroglyphA219,
    AnatolianHieroglyphA220,
    AnatolianHieroglyphA221,
    AnatolianHieroglyphA222,
    AnatolianHieroglyphA223,
    AnatolianHieroglyphA224,
    AnatolianHieroglyphA225,
    AnatolianHieroglyphA226,
    AnatolianHieroglyphA227,
    AnatolianHieroglyphA227a,
    AnatolianHieroglyphA228,
    AnatolianHieroglyphA229,
    AnatolianHieroglyphA230,
    AnatolianHieroglyphA231,
    AnatolianHieroglyphA232,
    AnatolianHieroglyphA233,
    AnatolianHieroglyphA234,
    AnatolianHieroglyphA235,
    AnatolianHieroglyphA236,
    AnatolianHieroglyphA237,
    AnatolianHieroglyphA238,
    AnatolianHieroglyphA239,
    AnatolianHieroglyphA240,
    AnatolianHieroglyphA241,
    AnatolianHieroglyphA242,
    AnatolianHieroglyphA243,
    AnatolianHieroglyphA244,
    AnatolianHieroglyphA245,
    AnatolianHieroglyphA246,
    AnatolianHieroglyphA247,
    AnatolianHieroglyphA248,
    AnatolianHieroglyphA249,
    AnatolianHieroglyphA250,
    AnatolianHieroglyphA251,
    AnatolianHieroglyphA252,
    AnatolianHieroglyphA253,
    AnatolianHieroglyphA254,
    AnatolianHieroglyphA255,
    AnatolianHieroglyphA256,
    AnatolianHieroglyphA257,
    AnatolianHieroglyphA258,
    AnatolianHieroglyphA259,
    AnatolianHieroglyphA260,
    AnatolianHieroglyphA261,
    AnatolianHieroglyphA262,
    AnatolianHieroglyphA263,
    AnatolianHieroglyphA264,
    AnatolianHieroglyphA265,
    AnatolianHieroglyphA266,
    AnatolianHieroglyphA267,
    AnatolianHieroglyphA267a,
    AnatolianHieroglyphA268,
    AnatolianHieroglyphA269,
    AnatolianHieroglyphA270,
    AnatolianHieroglyphA271,
    AnatolianHieroglyphA272,
    AnatolianHieroglyphA273,
    AnatolianHieroglyphA274,
    AnatolianHieroglyphA275,
    AnatolianHieroglyphA276,
    AnatolianHieroglyphA277,
    AnatolianHieroglyphA278,
    AnatolianHieroglyphA279,
    AnatolianHieroglyphA280,
    AnatolianHieroglyphA281,
    AnatolianHieroglyphA282,
    AnatolianHieroglyphA283,
    AnatolianHieroglyphA284,
    AnatolianHieroglyphA285,
    AnatolianHieroglyphA286,
    AnatolianHieroglyphA287,
    AnatolianHieroglyphA288,
    AnatolianHieroglyphA289,
    AnatolianHieroglyphA289a,
    AnatolianHieroglyphA290,
    AnatolianHieroglyphA291,
    AnatolianHieroglyphA292,
    AnatolianHieroglyphA293,
    AnatolianHieroglyphA294,
    AnatolianHieroglyphA294a,
    AnatolianHieroglyphA295,
    AnatolianHieroglyphA296,
    AnatolianHieroglyphA297,
    AnatolianHieroglyphA298,
    AnatolianHieroglyphA299,
    AnatolianHieroglyphA299a,
    AnatolianHieroglyphA300,
    AnatolianHieroglyphA301,
    AnatolianHieroglyphA302,
    AnatolianHieroglyphA303,
    AnatolianHieroglyphA304,
    AnatolianHieroglyphA305,
    AnatolianHieroglyphA306,
    AnatolianHieroglyphA307,
    AnatolianHieroglyphA308,
    AnatolianHieroglyphA309,
    AnatolianHieroglyphA309a,
    AnatolianHieroglyphA310,
    AnatolianHieroglyphA311,
    AnatolianHieroglyphA312,
    AnatolianHieroglyphA313,
    AnatolianHieroglyphA314,
    AnatolianHieroglyphA315,
    AnatolianHieroglyphA316,
    AnatolianHieroglyphA317,
    AnatolianHieroglyphA318,
    AnatolianHieroglyphA319,
    AnatolianHieroglyphA320,
    AnatolianHieroglyphA321,
    AnatolianHieroglyphA322,
    AnatolianHieroglyphA323,
    AnatolianHieroglyphA324,
    AnatolianHieroglyphA325,
    AnatolianHieroglyphA326,
    AnatolianHieroglyphA327,
    AnatolianHieroglyphA328,
    AnatolianHieroglyphA329,
    AnatolianHieroglyphA329a,
    AnatolianHieroglyphA330,
    AnatolianHieroglyphA331,
    AnatolianHieroglyphA332a,
    AnatolianHieroglyphA332b,
    AnatolianHieroglyphA332c,
    AnatolianHieroglyphA333,
    AnatolianHieroglyphA334,
    AnatolianHieroglyphA335,
    AnatolianHieroglyphA336,
    AnatolianHieroglyphA336a,
    AnatolianHieroglyphA336b,
    AnatolianHieroglyphA336c,
    AnatolianHieroglyphA337,
    AnatolianHieroglyphA338,
    AnatolianHieroglyphA339,
    AnatolianHieroglyphA340,
    AnatolianHieroglyphA341,
    AnatolianHieroglyphA342,
    AnatolianHieroglyphA343,
    AnatolianHieroglyphA344,
    AnatolianHieroglyphA345,
    AnatolianHieroglyphA346,
    AnatolianHieroglyphA347,
    AnatolianHieroglyphA348,
    AnatolianHieroglyphA349,
    AnatolianHieroglyphA350,
    AnatolianHieroglyphA351,
    AnatolianHieroglyphA352,
    AnatolianHieroglyphA353,
    AnatolianHieroglyphA354,
    AnatolianHieroglyphA355,
    AnatolianHieroglyphA356,
    AnatolianHieroglyphA357,
    AnatolianHieroglyphA358,
    AnatolianHieroglyphA359,
    AnatolianHieroglyphA359a,
    AnatolianHieroglyphA360,
    AnatolianHieroglyphA361,
    AnatolianHieroglyphA362,
    AnatolianHieroglyphA363,
    AnatolianHieroglyphA364,
    AnatolianHieroglyphA364a,
    AnatolianHieroglyphA365,
    AnatolianHieroglyphA366,
    AnatolianHieroglyphA367,
    AnatolianHieroglyphA368,
    AnatolianHieroglyphA368a,
    AnatolianHieroglyphA369,
    AnatolianHieroglyphA370,
    AnatolianHieroglyphA371,
    AnatolianHieroglyphA371a,
    AnatolianHieroglyphA372,
    AnatolianHieroglyphA373,
    AnatolianHieroglyphA374,
    AnatolianHieroglyphA375,
    AnatolianHieroglyphA376,
    AnatolianHieroglyphA377,
    AnatolianHieroglyphA378,
    AnatolianHieroglyphA379,
    AnatolianHieroglyphA380,
    AnatolianHieroglyphA381,
    AnatolianHieroglyphA381a,
    AnatolianHieroglyphA382,
    AnatolianHieroglyphA383RaOrRi,
    AnatolianHieroglyphA383a,
    AnatolianHieroglyphA384,
    AnatolianHieroglyphA385,
    AnatolianHieroglyphA386,
    AnatolianHieroglyphA386a,
    AnatolianHieroglyphA387,
    AnatolianHieroglyphA388,
    AnatolianHieroglyphA389,
    AnatolianHieroglyphA390,
    AnatolianHieroglyphA391,
    AnatolianHieroglyphA392,
    AnatolianHieroglyphA393Eight,
    AnatolianHieroglyphA394,
    AnatolianHieroglyphA395,
    AnatolianHieroglyphA396,
    AnatolianHieroglyphA397,
    AnatolianHieroglyphA398,
    AnatolianHieroglyphA399,
    AnatolianHieroglyphA400,
    AnatolianHieroglyphA401,
    AnatolianHieroglyphA402,
    AnatolianHieroglyphA403,
    AnatolianHieroglyphA404,
    AnatolianHieroglyphA405,
    AnatolianHieroglyphA406,
    AnatolianHieroglyphA407,
    AnatolianHieroglyphA408,
    AnatolianHieroglyphA409,
    AnatolianHieroglyphA410BeginLogogramMark,
    AnatolianHieroglyphA410aEndLogogramMark,
    AnatolianHieroglyphA411,
    AnatolianHieroglyphA412,
    AnatolianHieroglyphA413,
    AnatolianHieroglyphA414,
    AnatolianHieroglyphA415,
    AnatolianHieroglyphA416,
    AnatolianHieroglyphA417,
    AnatolianHieroglyphA418,
    AnatolianHieroglyphA419,
    AnatolianHieroglyphA420,
    AnatolianHieroglyphA421,
    AnatolianHieroglyphA422,
    AnatolianHieroglyphA423,
    AnatolianHieroglyphA424,
    AnatolianHieroglyphA425,
    AnatolianHieroglyphA426,
    AnatolianHieroglyphA427,
    AnatolianHieroglyphA428,
    AnatolianHieroglyphA429,
    AnatolianHieroglyphA430,
    AnatolianHieroglyphA431,
    AnatolianHieroglyphA432,
    AnatolianHieroglyphA433,
    AnatolianHieroglyphA434,
    AnatolianHieroglyphA435,
    AnatolianHieroglyphA436,
    AnatolianHieroglyphA437,
    AnatolianHieroglyphA438,
    AnatolianHieroglyphA439,
    AnatolianHieroglyphA440,
    AnatolianHieroglyphA441,
    AnatolianHieroglyphA442,
    AnatolianHieroglyphA443,
    AnatolianHieroglyphA444,
    AnatolianHieroglyphA445,
    AnatolianHieroglyphA446,
    AnatolianHieroglyphA447,
    AnatolianHieroglyphA448,
    AnatolianHieroglyphA449,
    AnatolianHieroglyphA450,
    AnatolianHieroglyphA450a,
    AnatolianHieroglyphA451,
    AnatolianHieroglyphA452,
    AnatolianHieroglyphA453,
    AnatolianHieroglyphA454,
    AnatolianHieroglyphA455,
    AnatolianHieroglyphA456,
    AnatolianHieroglyphA457,
    AnatolianHieroglyphA457a,
    AnatolianHieroglyphA458,
    AnatolianHieroglyphA459,
    AnatolianHieroglyphA460,
    AnatolianHieroglyphA461,
    AnatolianHieroglyphA462,
    AnatolianHieroglyphA463,
    AnatolianHieroglyphA464,
    AnatolianHieroglyphA465,
    AnatolianHieroglyphA466,
    AnatolianHieroglyphA467,
    AnatolianHieroglyphA468,
    AnatolianHieroglyphA469,
    AnatolianHieroglyphA470,
    AnatolianHieroglyphA471,
    AnatolianHieroglyphA472,
    AnatolianHieroglyphA473,
    AnatolianHieroglyphA474,
    AnatolianHieroglyphA475,
    AnatolianHieroglyphA476,
    AnatolianHieroglyphA477,
    AnatolianHieroglyphA478,
    AnatolianHieroglyphA479,
    AnatolianHieroglyphA480,
    AnatolianHieroglyphA481,
    AnatolianHieroglyphA482,
    AnatolianHieroglyphA483,
    AnatolianHieroglyphA484,
    AnatolianHieroglyphA485,
    AnatolianHieroglyphA486,
    AnatolianHieroglyphA487,
    AnatolianHieroglyphA488,
    AnatolianHieroglyphA489,
    AnatolianHieroglyphA490,
    AnatolianHieroglyphA491,
    AnatolianHieroglyphA492,
    AnatolianHieroglyphA493,
    AnatolianHieroglyphA494,
    AnatolianHieroglyphA495,
    AnatolianHieroglyphA496,
    AnatolianHieroglyphA497,
    AnatolianHieroglyphA501,
    AnatolianHieroglyphA502,
    AnatolianHieroglyphA503,
    AnatolianHieroglyphA504,
    AnatolianHieroglyphA505,
    AnatolianHieroglyphA506,
    AnatolianHieroglyphA507,
    AnatolianHieroglyphA508,
    AnatolianHieroglyphA509,
    AnatolianHieroglyphA510,
    AnatolianHieroglyphA511,
    AnatolianHieroglyphA512,
    AnatolianHieroglyphA513,
    AnatolianHieroglyphA514,
    AnatolianHieroglyphA515,
    AnatolianHieroglyphA516,
    AnatolianHieroglyphA517,
    AnatolianHieroglyphA518,
    AnatolianHieroglyphA519,
    AnatolianHieroglyphA520,
    AnatolianHieroglyphA521,
    AnatolianHieroglyphA522,
    AnatolianHieroglyphA523,
    AnatolianHieroglyphA524,
    AnatolianHieroglyphA525,
    AnatolianHieroglyphA526,
    AnatolianHieroglyphA527,
    AnatolianHieroglyphA528,
    AnatolianHieroglyphA529,
    AnatolianHieroglyphA530,
}

An enum to represent all characters in the AnatolianHieroglyphs block.

Variants
AnatolianHieroglyphA001

\u{14400}: '𔐀'

AnatolianHieroglyphA002

\u{14401}: '𔐁'

AnatolianHieroglyphA003

\u{14402}: '𔐂'

AnatolianHieroglyphA004

\u{14403}: '𔐃'

AnatolianHieroglyphA005

\u{14404}: '𔐄'

AnatolianHieroglyphA006

\u{14405}: '𔐅'

AnatolianHieroglyphA007

\u{14406}: '𔐆'

AnatolianHieroglyphA008

\u{14407}: '𔐇'

AnatolianHieroglyphA009

\u{14408}: '𔐈'

AnatolianHieroglyphA010

\u{14409}: '𔐉'

AnatolianHieroglyphA010a

\u{1440a}: '𔐊'

AnatolianHieroglyphA011

\u{1440b}: '𔐋'

AnatolianHieroglyphA012

\u{1440c}: '𔐌'

AnatolianHieroglyphA013

\u{1440d}: '𔐍'

AnatolianHieroglyphA014

\u{1440e}: '𔐎'

AnatolianHieroglyphA015

\u{1440f}: '𔐏'

AnatolianHieroglyphA016

\u{14410}: '𔐐'

AnatolianHieroglyphA017

\u{14411}: '𔐑'

AnatolianHieroglyphA018

\u{14412}: '𔐒'

AnatolianHieroglyphA019

\u{14413}: '𔐓'

AnatolianHieroglyphA020

\u{14414}: '𔐔'

AnatolianHieroglyphA021

\u{14415}: '𔐕'

AnatolianHieroglyphA022

\u{14416}: '𔐖'

AnatolianHieroglyphA023

\u{14417}: '𔐗'

AnatolianHieroglyphA024

\u{14418}: '𔐘'

AnatolianHieroglyphA025

\u{14419}: '𔐙'

AnatolianHieroglyphA026

\u{1441a}: '𔐚'

AnatolianHieroglyphA026a

\u{1441b}: '𔐛'

AnatolianHieroglyphA027

\u{1441c}: '𔐜'

AnatolianHieroglyphA028

\u{1441d}: '𔐝'

AnatolianHieroglyphA029

\u{1441e}: '𔐞'

AnatolianHieroglyphA030

\u{1441f}: '𔐟'

AnatolianHieroglyphA031

\u{14420}: '𔐠'

AnatolianHieroglyphA032

\u{14421}: '𔐡'

AnatolianHieroglyphA033

\u{14422}: '𔐢'

AnatolianHieroglyphA034

\u{14423}: '𔐣'

AnatolianHieroglyphA035

\u{14424}: '𔐤'

AnatolianHieroglyphA036

\u{14425}: '𔐥'

AnatolianHieroglyphA037

\u{14426}: '𔐦'

AnatolianHieroglyphA038

\u{14427}: '𔐧'

AnatolianHieroglyphA039

\u{14428}: '𔐨'

AnatolianHieroglyphA039a

\u{14429}: '𔐩'

AnatolianHieroglyphA040

\u{1442a}: '𔐪'

AnatolianHieroglyphA041

\u{1442b}: '𔐫'

AnatolianHieroglyphA041a

\u{1442c}: '𔐬'

AnatolianHieroglyphA042

\u{1442d}: '𔐭'

AnatolianHieroglyphA043

\u{1442e}: '𔐮'

AnatolianHieroglyphA044

\u{1442f}: '𔐯'

AnatolianHieroglyphA045

\u{14430}: '𔐰'

AnatolianHieroglyphA045a

\u{14431}: '𔐱'

AnatolianHieroglyphA046

\u{14432}: '𔐲'

AnatolianHieroglyphA046a

\u{14433}: '𔐳'

AnatolianHieroglyphA046b

\u{14434}: '𔐴'

AnatolianHieroglyphA047

\u{14435}: '𔐵'

AnatolianHieroglyphA048

\u{14436}: '𔐶'

AnatolianHieroglyphA049

\u{14437}: '𔐷'

AnatolianHieroglyphA050

\u{14438}: '𔐸'

AnatolianHieroglyphA051

\u{14439}: '𔐹'

AnatolianHieroglyphA052

\u{1443a}: '𔐺'

AnatolianHieroglyphA053

\u{1443b}: '𔐻'

AnatolianHieroglyphA054

\u{1443c}: '𔐼'

AnatolianHieroglyphA055

\u{1443d}: '𔐽'

AnatolianHieroglyphA056

\u{1443e}: '𔐾'

AnatolianHieroglyphA057

\u{1443f}: '𔐿'

AnatolianHieroglyphA058

\u{14440}: '𔑀'

AnatolianHieroglyphA059

\u{14441}: '𔑁'

AnatolianHieroglyphA060

\u{14442}: '𔑂'

AnatolianHieroglyphA061

\u{14443}: '𔑃'

AnatolianHieroglyphA062

\u{14444}: '𔑄'

AnatolianHieroglyphA063

\u{14445}: '𔑅'

AnatolianHieroglyphA064

\u{14446}: '𔑆'

AnatolianHieroglyphA065

\u{14447}: '𔑇'

AnatolianHieroglyphA066

\u{14448}: '𔑈'

AnatolianHieroglyphA066a

\u{14449}: '𔑉'

AnatolianHieroglyphA066b

\u{1444a}: '𔑊'

AnatolianHieroglyphA066c

\u{1444b}: '𔑋'

AnatolianHieroglyphA067

\u{1444c}: '𔑌'

AnatolianHieroglyphA068

\u{1444d}: '𔑍'

AnatolianHieroglyphA069

\u{1444e}: '𔑎'

AnatolianHieroglyphA070

\u{1444f}: '𔑏'

AnatolianHieroglyphA071

\u{14450}: '𔑐'

AnatolianHieroglyphA072

\u{14451}: '𔑑'

AnatolianHieroglyphA073

\u{14452}: '𔑒'

AnatolianHieroglyphA074

\u{14453}: '𔑓'

AnatolianHieroglyphA075

\u{14454}: '𔑔'

AnatolianHieroglyphA076

\u{14455}: '𔑕'

AnatolianHieroglyphA077

\u{14456}: '𔑖'

AnatolianHieroglyphA078

\u{14457}: '𔑗'

AnatolianHieroglyphA079

\u{14458}: '𔑘'

AnatolianHieroglyphA080

\u{14459}: '𔑙'

AnatolianHieroglyphA081

\u{1445a}: '𔑚'

AnatolianHieroglyphA082

\u{1445b}: '𔑛'

AnatolianHieroglyphA083

\u{1445c}: '𔑜'

AnatolianHieroglyphA084

\u{1445d}: '𔑝'

AnatolianHieroglyphA085

\u{1445e}: '𔑞'

AnatolianHieroglyphA086

\u{1445f}: '𔑟'

AnatolianHieroglyphA087

\u{14460}: '𔑠'

AnatolianHieroglyphA088

\u{14461}: '𔑡'

AnatolianHieroglyphA089

\u{14462}: '𔑢'

AnatolianHieroglyphA090

\u{14463}: '𔑣'

AnatolianHieroglyphA091

\u{14464}: '𔑤'

AnatolianHieroglyphA092

\u{14465}: '𔑥'

AnatolianHieroglyphA093

\u{14466}: '𔑦'

AnatolianHieroglyphA094

\u{14467}: '𔑧'

AnatolianHieroglyphA095

\u{14468}: '𔑨'

AnatolianHieroglyphA096

\u{14469}: '𔑩'

AnatolianHieroglyphA097

\u{1446a}: '𔑪'

AnatolianHieroglyphA097a

\u{1446b}: '𔑫'

AnatolianHieroglyphA098

\u{1446c}: '𔑬'

AnatolianHieroglyphA098a

\u{1446d}: '𔑭'

AnatolianHieroglyphA099

\u{1446e}: '𔑮'

AnatolianHieroglyphA100

\u{1446f}: '𔑯'

AnatolianHieroglyphA100a

\u{14470}: '𔑰'

AnatolianHieroglyphA101

\u{14471}: '𔑱'

AnatolianHieroglyphA101a

\u{14472}: '𔑲'

AnatolianHieroglyphA102

\u{14473}: '𔑳'

AnatolianHieroglyphA102a

\u{14474}: '𔑴'

AnatolianHieroglyphA103

\u{14475}: '𔑵'

AnatolianHieroglyphA104

\u{14476}: '𔑶'

AnatolianHieroglyphA104a

\u{14477}: '𔑷'

AnatolianHieroglyphA104b

\u{14478}: '𔑸'

AnatolianHieroglyphA104c

\u{14479}: '𔑹'

AnatolianHieroglyphA105

\u{1447a}: '𔑺'

AnatolianHieroglyphA105a

\u{1447b}: '𔑻'

AnatolianHieroglyphA105b

\u{1447c}: '𔑼'

AnatolianHieroglyphA106

\u{1447d}: '𔑽'

AnatolianHieroglyphA107

\u{1447e}: '𔑾'

AnatolianHieroglyphA107a

\u{1447f}: '𔑿'

AnatolianHieroglyphA107b

\u{14480}: '𔒀'

AnatolianHieroglyphA107c

\u{14481}: '𔒁'

AnatolianHieroglyphA108

\u{14482}: '𔒂'

AnatolianHieroglyphA109

\u{14483}: '𔒃'

AnatolianHieroglyphA110

\u{14484}: '𔒄'

AnatolianHieroglyphA110a

\u{14485}: '𔒅'

AnatolianHieroglyphA110b

\u{14486}: '𔒆'

AnatolianHieroglyphA111

\u{14487}: '𔒇'

AnatolianHieroglyphA112

\u{14488}: '𔒈'

AnatolianHieroglyphA113

\u{14489}: '𔒉'

AnatolianHieroglyphA114

\u{1448a}: '𔒊'

AnatolianHieroglyphA115

\u{1448b}: '𔒋'

AnatolianHieroglyphA115a

\u{1448c}: '𔒌'

AnatolianHieroglyphA116

\u{1448d}: '𔒍'

AnatolianHieroglyphA117

\u{1448e}: '𔒎'

AnatolianHieroglyphA118

\u{1448f}: '𔒏'

AnatolianHieroglyphA119

\u{14490}: '𔒐'

AnatolianHieroglyphA120

\u{14491}: '𔒑'

AnatolianHieroglyphA121

\u{14492}: '𔒒'

AnatolianHieroglyphA122

\u{14493}: '𔒓'

AnatolianHieroglyphA123

\u{14494}: '𔒔'

AnatolianHieroglyphA124

\u{14495}: '𔒕'

AnatolianHieroglyphA125

\u{14496}: '𔒖'

AnatolianHieroglyphA125a

\u{14497}: '𔒗'

AnatolianHieroglyphA126

\u{14498}: '𔒘'

AnatolianHieroglyphA127

\u{14499}: '𔒙'

AnatolianHieroglyphA128

\u{1449a}: '𔒚'

AnatolianHieroglyphA129

\u{1449b}: '𔒛'

AnatolianHieroglyphA130

\u{1449c}: '𔒜'

AnatolianHieroglyphA131

\u{1449d}: '𔒝'

AnatolianHieroglyphA132

\u{1449e}: '𔒞'

AnatolianHieroglyphA133

\u{1449f}: '𔒟'

AnatolianHieroglyphA134

\u{144a0}: '𔒠'

AnatolianHieroglyphA135

\u{144a1}: '𔒡'

AnatolianHieroglyphA135a

\u{144a2}: '𔒢'

AnatolianHieroglyphA136

\u{144a3}: '𔒣'

AnatolianHieroglyphA137

\u{144a4}: '𔒤'

AnatolianHieroglyphA138

\u{144a5}: '𔒥'

AnatolianHieroglyphA139

\u{144a6}: '𔒦'

AnatolianHieroglyphA140

\u{144a7}: '𔒧'

AnatolianHieroglyphA141

\u{144a8}: '𔒨'

AnatolianHieroglyphA142

\u{144a9}: '𔒩'

AnatolianHieroglyphA143

\u{144aa}: '𔒪'

AnatolianHieroglyphA144

\u{144ab}: '𔒫'

AnatolianHieroglyphA145

\u{144ac}: '𔒬'

AnatolianHieroglyphA146

\u{144ad}: '𔒭'

AnatolianHieroglyphA147

\u{144ae}: '𔒮'

AnatolianHieroglyphA148

\u{144af}: '𔒯'

AnatolianHieroglyphA149

\u{144b0}: '𔒰'

AnatolianHieroglyphA150

\u{144b1}: '𔒱'

AnatolianHieroglyphA151

\u{144b2}: '𔒲'

AnatolianHieroglyphA152

\u{144b3}: '𔒳'

AnatolianHieroglyphA153

\u{144b4}: '𔒴'

AnatolianHieroglyphA154

\u{144b5}: '𔒵'

AnatolianHieroglyphA155

\u{144b6}: '𔒶'

AnatolianHieroglyphA156

\u{144b7}: '𔒷'

AnatolianHieroglyphA157

\u{144b8}: '𔒸'

AnatolianHieroglyphA158

\u{144b9}: '𔒹'

AnatolianHieroglyphA159

\u{144ba}: '𔒺'

AnatolianHieroglyphA160

\u{144bb}: '𔒻'

AnatolianHieroglyphA161

\u{144bc}: '𔒼'

AnatolianHieroglyphA162

\u{144bd}: '𔒽'

AnatolianHieroglyphA163

\u{144be}: '𔒾'

AnatolianHieroglyphA164

\u{144bf}: '𔒿'

AnatolianHieroglyphA165

\u{144c0}: '𔓀'

AnatolianHieroglyphA166

\u{144c1}: '𔓁'

AnatolianHieroglyphA167

\u{144c2}: '𔓂'

AnatolianHieroglyphA168

\u{144c3}: '𔓃'

AnatolianHieroglyphA169

\u{144c4}: '𔓄'

AnatolianHieroglyphA170

\u{144c5}: '𔓅'

AnatolianHieroglyphA171

\u{144c6}: '𔓆'

AnatolianHieroglyphA172

\u{144c7}: '𔓇'

AnatolianHieroglyphA173

\u{144c8}: '𔓈'

AnatolianHieroglyphA174

\u{144c9}: '𔓉'

AnatolianHieroglyphA175

\u{144ca}: '𔓊'

AnatolianHieroglyphA176

\u{144cb}: '𔓋'

AnatolianHieroglyphA177

\u{144cc}: '𔓌'

AnatolianHieroglyphA178

\u{144cd}: '𔓍'

AnatolianHieroglyphA179

\u{144ce}: '𔓎'

AnatolianHieroglyphA180

\u{144cf}: '𔓏'

AnatolianHieroglyphA181

\u{144d0}: '𔓐'

AnatolianHieroglyphA182

\u{144d1}: '𔓑'

AnatolianHieroglyphA183

\u{144d2}: '𔓒'

AnatolianHieroglyphA184

\u{144d3}: '𔓓'

AnatolianHieroglyphA185

\u{144d4}: '𔓔'

AnatolianHieroglyphA186

\u{144d5}: '𔓕'

AnatolianHieroglyphA187

\u{144d6}: '𔓖'

AnatolianHieroglyphA188

\u{144d7}: '𔓗'

AnatolianHieroglyphA189

\u{144d8}: '𔓘'

AnatolianHieroglyphA190

\u{144d9}: '𔓙'

AnatolianHieroglyphA191

\u{144da}: '𔓚'

AnatolianHieroglyphA192

\u{144db}: '𔓛'

AnatolianHieroglyphA193

\u{144dc}: '𔓜'

AnatolianHieroglyphA194

\u{144dd}: '𔓝'

AnatolianHieroglyphA195

\u{144de}: '𔓞'

AnatolianHieroglyphA196

\u{144df}: '𔓟'

AnatolianHieroglyphA197

\u{144e0}: '𔓠'

AnatolianHieroglyphA198

\u{144e1}: '𔓡'

AnatolianHieroglyphA199

\u{144e2}: '𔓢'

AnatolianHieroglyphA200

\u{144e3}: '𔓣'

AnatolianHieroglyphA201

\u{144e4}: '𔓤'

AnatolianHieroglyphA202

\u{144e5}: '𔓥'

AnatolianHieroglyphA202a

\u{144e6}: '𔓦'

AnatolianHieroglyphA202b

\u{144e7}: '𔓧'

AnatolianHieroglyphA203

\u{144e8}: '𔓨'

AnatolianHieroglyphA204

\u{144e9}: '𔓩'

AnatolianHieroglyphA205

\u{144ea}: '𔓪'

AnatolianHieroglyphA206

\u{144eb}: '𔓫'

AnatolianHieroglyphA207

\u{144ec}: '𔓬'

AnatolianHieroglyphA207a

\u{144ed}: '𔓭'

AnatolianHieroglyphA208

\u{144ee}: '𔓮'

AnatolianHieroglyphA209

\u{144ef}: '𔓯'

AnatolianHieroglyphA209a

\u{144f0}: '𔓰'

AnatolianHieroglyphA210

\u{144f1}: '𔓱'

AnatolianHieroglyphA211

\u{144f2}: '𔓲'

AnatolianHieroglyphA212

\u{144f3}: '𔓳'

AnatolianHieroglyphA213

\u{144f4}: '𔓴'

AnatolianHieroglyphA214

\u{144f5}: '𔓵'

AnatolianHieroglyphA215

\u{144f6}: '𔓶'

AnatolianHieroglyphA215a

\u{144f7}: '𔓷'

AnatolianHieroglyphA216

\u{144f8}: '𔓸'

AnatolianHieroglyphA216a

\u{144f9}: '𔓹'

AnatolianHieroglyphA217

\u{144fa}: '𔓺'

AnatolianHieroglyphA218

\u{144fb}: '𔓻'

AnatolianHieroglyphA219

\u{144fc}: '𔓼'

AnatolianHieroglyphA220

\u{144fd}: '𔓽'

AnatolianHieroglyphA221

\u{144fe}: '𔓾'

AnatolianHieroglyphA222

\u{144ff}: '𔓿'

AnatolianHieroglyphA223

\u{14500}: '𔔀'

AnatolianHieroglyphA224

\u{14501}: '𔔁'

AnatolianHieroglyphA225

\u{14502}: '𔔂'

AnatolianHieroglyphA226

\u{14503}: '𔔃'

AnatolianHieroglyphA227

\u{14504}: '𔔄'

AnatolianHieroglyphA227a

\u{14505}: '𔔅'

AnatolianHieroglyphA228

\u{14506}: '𔔆'

AnatolianHieroglyphA229

\u{14507}: '𔔇'

AnatolianHieroglyphA230

\u{14508}: '𔔈'

AnatolianHieroglyphA231

\u{14509}: '𔔉'

AnatolianHieroglyphA232

\u{1450a}: '𔔊'

AnatolianHieroglyphA233

\u{1450b}: '𔔋'

AnatolianHieroglyphA234

\u{1450c}: '𔔌'

AnatolianHieroglyphA235

\u{1450d}: '𔔍'

AnatolianHieroglyphA236

\u{1450e}: '𔔎'

AnatolianHieroglyphA237

\u{1450f}: '𔔏'

AnatolianHieroglyphA238

\u{14510}: '𔔐'

AnatolianHieroglyphA239

\u{14511}: '𔔑'

AnatolianHieroglyphA240

\u{14512}: '𔔒'

AnatolianHieroglyphA241

\u{14513}: '𔔓'

AnatolianHieroglyphA242

\u{14514}: '𔔔'

AnatolianHieroglyphA243

\u{14515}: '𔔕'

AnatolianHieroglyphA244

\u{14516}: '𔔖'

AnatolianHieroglyphA245

\u{14517}: '𔔗'

AnatolianHieroglyphA246

\u{14518}: '𔔘'

AnatolianHieroglyphA247

\u{14519}: '𔔙'

AnatolianHieroglyphA248

\u{1451a}: '𔔚'

AnatolianHieroglyphA249

\u{1451b}: '𔔛'

AnatolianHieroglyphA250

\u{1451c}: '𔔜'

AnatolianHieroglyphA251

\u{1451d}: '𔔝'

AnatolianHieroglyphA252

\u{1451e}: '𔔞'

AnatolianHieroglyphA253

\u{1451f}: '𔔟'

AnatolianHieroglyphA254

\u{14520}: '𔔠'

AnatolianHieroglyphA255

\u{14521}: '𔔡'

AnatolianHieroglyphA256

\u{14522}: '𔔢'

AnatolianHieroglyphA257

\u{14523}: '𔔣'

AnatolianHieroglyphA258

\u{14524}: '𔔤'

AnatolianHieroglyphA259

\u{14525}: '𔔥'

AnatolianHieroglyphA260

\u{14526}: '𔔦'

AnatolianHieroglyphA261

\u{14527}: '𔔧'

AnatolianHieroglyphA262

\u{14528}: '𔔨'

AnatolianHieroglyphA263

\u{14529}: '𔔩'

AnatolianHieroglyphA264

\u{1452a}: '𔔪'

AnatolianHieroglyphA265

\u{1452b}: '𔔫'

AnatolianHieroglyphA266

\u{1452c}: '𔔬'

AnatolianHieroglyphA267

\u{1452d}: '𔔭'

AnatolianHieroglyphA267a

\u{1452e}: '𔔮'

AnatolianHieroglyphA268

\u{1452f}: '𔔯'

AnatolianHieroglyphA269

\u{14530}: '𔔰'

AnatolianHieroglyphA270

\u{14531}: '𔔱'

AnatolianHieroglyphA271

\u{14532}: '𔔲'

AnatolianHieroglyphA272

\u{14533}: '𔔳'

AnatolianHieroglyphA273

\u{14534}: '𔔴'

AnatolianHieroglyphA274

\u{14535}: '𔔵'

AnatolianHieroglyphA275

\u{14536}: '𔔶'

AnatolianHieroglyphA276

\u{14537}: '𔔷'

AnatolianHieroglyphA277

\u{14538}: '𔔸'

AnatolianHieroglyphA278

\u{14539}: '𔔹'

AnatolianHieroglyphA279

\u{1453a}: '𔔺'

AnatolianHieroglyphA280

\u{1453b}: '𔔻'

AnatolianHieroglyphA281

\u{1453c}: '𔔼'

AnatolianHieroglyphA282

\u{1453d}: '𔔽'

AnatolianHieroglyphA283

\u{1453e}: '𔔾'

AnatolianHieroglyphA284

\u{1453f}: '𔔿'

AnatolianHieroglyphA285

\u{14540}: '𔕀'

AnatolianHieroglyphA286

\u{14541}: '𔕁'

AnatolianHieroglyphA287

\u{14542}: '𔕂'

AnatolianHieroglyphA288

\u{14543}: '𔕃'

AnatolianHieroglyphA289

\u{14544}: '𔕄'

AnatolianHieroglyphA289a

\u{14545}: '𔕅'

AnatolianHieroglyphA290

\u{14546}: '𔕆'

AnatolianHieroglyphA291

\u{14547}: '𔕇'

AnatolianHieroglyphA292

\u{14548}: '𔕈'

AnatolianHieroglyphA293

\u{14549}: '𔕉'

AnatolianHieroglyphA294

\u{1454a}: '𔕊'

AnatolianHieroglyphA294a

\u{1454b}: '𔕋'

AnatolianHieroglyphA295

\u{1454c}: '𔕌'

AnatolianHieroglyphA296

\u{1454d}: '𔕍'

AnatolianHieroglyphA297

\u{1454e}: '𔕎'

AnatolianHieroglyphA298

\u{1454f}: '𔕏'

AnatolianHieroglyphA299

\u{14550}: '𔕐'

AnatolianHieroglyphA299a

\u{14551}: '𔕑'

AnatolianHieroglyphA300

\u{14552}: '𔕒'

AnatolianHieroglyphA301

\u{14553}: '𔕓'

AnatolianHieroglyphA302

\u{14554}: '𔕔'

AnatolianHieroglyphA303

\u{14555}: '𔕕'

AnatolianHieroglyphA304

\u{14556}: '𔕖'

AnatolianHieroglyphA305

\u{14557}: '𔕗'

AnatolianHieroglyphA306

\u{14558}: '𔕘'

AnatolianHieroglyphA307

\u{14559}: '𔕙'

AnatolianHieroglyphA308

\u{1455a}: '𔕚'

AnatolianHieroglyphA309

\u{1455b}: '𔕛'

AnatolianHieroglyphA309a

\u{1455c}: '𔕜'

AnatolianHieroglyphA310

\u{1455d}: '𔕝'

AnatolianHieroglyphA311

\u{1455e}: '𔕞'

AnatolianHieroglyphA312

\u{1455f}: '𔕟'

AnatolianHieroglyphA313

\u{14560}: '𔕠'

AnatolianHieroglyphA314

\u{14561}: '𔕡'

AnatolianHieroglyphA315

\u{14562}: '𔕢'

AnatolianHieroglyphA316

\u{14563}: '𔕣'

AnatolianHieroglyphA317

\u{14564}: '𔕤'

AnatolianHieroglyphA318

\u{14565}: '𔕥'

AnatolianHieroglyphA319

\u{14566}: '𔕦'

AnatolianHieroglyphA320

\u{14567}: '𔕧'

AnatolianHieroglyphA321

\u{14568}: '𔕨'

AnatolianHieroglyphA322

\u{14569}: '𔕩'

AnatolianHieroglyphA323

\u{1456a}: '𔕪'

AnatolianHieroglyphA324

\u{1456b}: '𔕫'

AnatolianHieroglyphA325

\u{1456c}: '𔕬'

AnatolianHieroglyphA326

\u{1456d}: '𔕭'

AnatolianHieroglyphA327

\u{1456e}: '𔕮'

AnatolianHieroglyphA328

\u{1456f}: '𔕯'

AnatolianHieroglyphA329

\u{14570}: '𔕰'

AnatolianHieroglyphA329a

\u{14571}: '𔕱'

AnatolianHieroglyphA330

\u{14572}: '𔕲'

AnatolianHieroglyphA331

\u{14573}: '𔕳'

AnatolianHieroglyphA332a

\u{14574}: '𔕴'

AnatolianHieroglyphA332b

\u{14575}: '𔕵'

AnatolianHieroglyphA332c

\u{14576}: '𔕶'

AnatolianHieroglyphA333

\u{14577}: '𔕷'

AnatolianHieroglyphA334

\u{14578}: '𔕸'

AnatolianHieroglyphA335

\u{14579}: '𔕹'

AnatolianHieroglyphA336

\u{1457a}: '𔕺'

AnatolianHieroglyphA336a

\u{1457b}: '𔕻'

AnatolianHieroglyphA336b

\u{1457c}: '𔕼'

AnatolianHieroglyphA336c

\u{1457d}: '𔕽'

AnatolianHieroglyphA337

\u{1457e}: '𔕾'

AnatolianHieroglyphA338

\u{1457f}: '𔕿'

AnatolianHieroglyphA339

\u{14580}: '𔖀'

AnatolianHieroglyphA340

\u{14581}: '𔖁'

AnatolianHieroglyphA341

\u{14582}: '𔖂'

AnatolianHieroglyphA342

\u{14583}: '𔖃'

AnatolianHieroglyphA343

\u{14584}: '𔖄'

AnatolianHieroglyphA344

\u{14585}: '𔖅'

AnatolianHieroglyphA345

\u{14586}: '𔖆'

AnatolianHieroglyphA346

\u{14587}: '𔖇'

AnatolianHieroglyphA347

\u{14588}: '𔖈'

AnatolianHieroglyphA348

\u{14589}: '𔖉'

AnatolianHieroglyphA349

\u{1458a}: '𔖊'

AnatolianHieroglyphA350

\u{1458b}: '𔖋'

AnatolianHieroglyphA351

\u{1458c}: '𔖌'

AnatolianHieroglyphA352

\u{1458d}: '𔖍'

AnatolianHieroglyphA353

\u{1458e}: '𔖎'

AnatolianHieroglyphA354

\u{1458f}: '𔖏'

AnatolianHieroglyphA355

\u{14590}: '𔖐'

AnatolianHieroglyphA356

\u{14591}: '𔖑'

AnatolianHieroglyphA357

\u{14592}: '𔖒'

AnatolianHieroglyphA358

\u{14593}: '𔖓'

AnatolianHieroglyphA359

\u{14594}: '𔖔'

AnatolianHieroglyphA359a

\u{14595}: '𔖕'

AnatolianHieroglyphA360

\u{14596}: '𔖖'

AnatolianHieroglyphA361

\u{14597}: '𔖗'

AnatolianHieroglyphA362

\u{14598}: '𔖘'

AnatolianHieroglyphA363

\u{14599}: '𔖙'

AnatolianHieroglyphA364

\u{1459a}: '𔖚'

AnatolianHieroglyphA364a

\u{1459b}: '𔖛'

AnatolianHieroglyphA365

\u{1459c}: '𔖜'

AnatolianHieroglyphA366

\u{1459d}: '𔖝'

AnatolianHieroglyphA367

\u{1459e}: '𔖞'

AnatolianHieroglyphA368

\u{1459f}: '𔖟'

AnatolianHieroglyphA368a

\u{145a0}: '𔖠'

AnatolianHieroglyphA369

\u{145a1}: '𔖡'

AnatolianHieroglyphA370

\u{145a2}: '𔖢'

AnatolianHieroglyphA371

\u{145a3}: '𔖣'

AnatolianHieroglyphA371a

\u{145a4}: '𔖤'

AnatolianHieroglyphA372

\u{145a5}: '𔖥'

AnatolianHieroglyphA373

\u{145a6}: '𔖦'

AnatolianHieroglyphA374

\u{145a7}: '𔖧'

AnatolianHieroglyphA375

\u{145a8}: '𔖨'

AnatolianHieroglyphA376

\u{145a9}: '𔖩'

AnatolianHieroglyphA377

\u{145aa}: '𔖪'

AnatolianHieroglyphA378

\u{145ab}: '𔖫'

AnatolianHieroglyphA379

\u{145ac}: '𔖬'

AnatolianHieroglyphA380

\u{145ad}: '𔖭'

AnatolianHieroglyphA381

\u{145ae}: '𔖮'

AnatolianHieroglyphA381a

\u{145af}: '𔖯'

AnatolianHieroglyphA382

\u{145b0}: '𔖰'

AnatolianHieroglyphA383RaOrRi

\u{145b1}: '𔖱'

AnatolianHieroglyphA383a

\u{145b2}: '𔖲'

AnatolianHieroglyphA384

\u{145b3}: '𔖳'

AnatolianHieroglyphA385

\u{145b4}: '𔖴'

AnatolianHieroglyphA386

\u{145b5}: '𔖵'

AnatolianHieroglyphA386a

\u{145b6}: '𔖶'

AnatolianHieroglyphA387

\u{145b7}: '𔖷'

AnatolianHieroglyphA388

\u{145b8}: '𔖸'

AnatolianHieroglyphA389

\u{145b9}: '𔖹'

AnatolianHieroglyphA390

\u{145ba}: '𔖺'

AnatolianHieroglyphA391

\u{145bb}: '𔖻'

AnatolianHieroglyphA392

\u{145bc}: '𔖼'

AnatolianHieroglyphA393Eight

\u{145bd}: '𔖽'

AnatolianHieroglyphA394

\u{145be}: '𔖾'

AnatolianHieroglyphA395

\u{145bf}: '𔖿'

AnatolianHieroglyphA396

\u{145c0}: '𔗀'

AnatolianHieroglyphA397

\u{145c1}: '𔗁'

AnatolianHieroglyphA398

\u{145c2}: '𔗂'

AnatolianHieroglyphA399

\u{145c3}: '𔗃'

AnatolianHieroglyphA400

\u{145c4}: '𔗄'

AnatolianHieroglyphA401

\u{145c5}: '𔗅'

AnatolianHieroglyphA402

\u{145c6}: '𔗆'

AnatolianHieroglyphA403

\u{145c7}: '𔗇'

AnatolianHieroglyphA404

\u{145c8}: '𔗈'

AnatolianHieroglyphA405

\u{145c9}: '𔗉'

AnatolianHieroglyphA406

\u{145ca}: '𔗊'

AnatolianHieroglyphA407

\u{145cb}: '𔗋'

AnatolianHieroglyphA408

\u{145cc}: '𔗌'

AnatolianHieroglyphA409

\u{145cd}: '𔗍'

AnatolianHieroglyphA410BeginLogogramMark

\u{145ce}: '𔗎'

AnatolianHieroglyphA410aEndLogogramMark

\u{145cf}: '𔗏'

AnatolianHieroglyphA411

\u{145d0}: '𔗐'

AnatolianHieroglyphA412

\u{145d1}: '𔗑'

AnatolianHieroglyphA413

\u{145d2}: '𔗒'

AnatolianHieroglyphA414

\u{145d3}: '𔗓'

AnatolianHieroglyphA415

\u{145d4}: '𔗔'

AnatolianHieroglyphA416

\u{145d5}: '𔗕'

AnatolianHieroglyphA417

\u{145d6}: '𔗖'

AnatolianHieroglyphA418

\u{145d7}: '𔗗'

AnatolianHieroglyphA419

\u{145d8}: '𔗘'

AnatolianHieroglyphA420

\u{145d9}: '𔗙'

AnatolianHieroglyphA421

\u{145da}: '𔗚'

AnatolianHieroglyphA422

\u{145db}: '𔗛'

AnatolianHieroglyphA423

\u{145dc}: '𔗜'

AnatolianHieroglyphA424

\u{145dd}: '𔗝'

AnatolianHieroglyphA425

\u{145de}: '𔗞'

AnatolianHieroglyphA426

\u{145df}: '𔗟'

AnatolianHieroglyphA427

\u{145e0}: '𔗠'

AnatolianHieroglyphA428

\u{145e1}: '𔗡'

AnatolianHieroglyphA429

\u{145e2}: '𔗢'

AnatolianHieroglyphA430

\u{145e3}: '𔗣'

AnatolianHieroglyphA431

\u{145e4}: '𔗤'

AnatolianHieroglyphA432

\u{145e5}: '𔗥'

AnatolianHieroglyphA433

\u{145e6}: '𔗦'

AnatolianHieroglyphA434

\u{145e7}: '𔗧'

AnatolianHieroglyphA435

\u{145e8}: '𔗨'

AnatolianHieroglyphA436

\u{145e9}: '𔗩'

AnatolianHieroglyphA437

\u{145ea}: '𔗪'

AnatolianHieroglyphA438

\u{145eb}: '𔗫'

AnatolianHieroglyphA439

\u{145ec}: '𔗬'

AnatolianHieroglyphA440

\u{145ed}: '𔗭'

AnatolianHieroglyphA441

\u{145ee}: '𔗮'

AnatolianHieroglyphA442

\u{145ef}: '𔗯'

AnatolianHieroglyphA443

\u{145f0}: '𔗰'

AnatolianHieroglyphA444

\u{145f1}: '𔗱'

AnatolianHieroglyphA445

\u{145f2}: '𔗲'

AnatolianHieroglyphA446

\u{145f3}: '𔗳'

AnatolianHieroglyphA447

\u{145f4}: '𔗴'

AnatolianHieroglyphA448

\u{145f5}: '𔗵'

AnatolianHieroglyphA449

\u{145f6}: '𔗶'

AnatolianHieroglyphA450

\u{145f7}: '𔗷'

AnatolianHieroglyphA450a

\u{145f8}: '𔗸'

AnatolianHieroglyphA451

\u{145f9}: '𔗹'

AnatolianHieroglyphA452

\u{145fa}: '𔗺'

AnatolianHieroglyphA453

\u{145fb}: '𔗻'

AnatolianHieroglyphA454

\u{145fc}: '𔗼'

AnatolianHieroglyphA455

\u{145fd}: '𔗽'

AnatolianHieroglyphA456

\u{145fe}: '𔗾'

AnatolianHieroglyphA457

\u{145ff}: '𔗿'

AnatolianHieroglyphA457a

\u{14600}: '𔘀'

AnatolianHieroglyphA458

\u{14601}: '𔘁'

AnatolianHieroglyphA459

\u{14602}: '𔘂'

AnatolianHieroglyphA460

\u{14603}: '𔘃'

AnatolianHieroglyphA461

\u{14604}: '𔘄'

AnatolianHieroglyphA462

\u{14605}: '𔘅'

AnatolianHieroglyphA463

\u{14606}: '𔘆'

AnatolianHieroglyphA464

\u{14607}: '𔘇'

AnatolianHieroglyphA465

\u{14608}: '𔘈'

AnatolianHieroglyphA466

\u{14609}: '𔘉'

AnatolianHieroglyphA467

\u{1460a}: '𔘊'

AnatolianHieroglyphA468

\u{1460b}: '𔘋'

AnatolianHieroglyphA469

\u{1460c}: '𔘌'

AnatolianHieroglyphA470

\u{1460d}: '𔘍'

AnatolianHieroglyphA471

\u{1460e}: '𔘎'

AnatolianHieroglyphA472

\u{1460f}: '𔘏'

AnatolianHieroglyphA473

\u{14610}: '𔘐'

AnatolianHieroglyphA474

\u{14611}: '𔘑'

AnatolianHieroglyphA475

\u{14612}: '𔘒'

AnatolianHieroglyphA476

\u{14613}: '𔘓'

AnatolianHieroglyphA477

\u{14614}: '𔘔'

AnatolianHieroglyphA478

\u{14615}: '𔘕'

AnatolianHieroglyphA479

\u{14616}: '𔘖'

AnatolianHieroglyphA480

\u{14617}: '𔘗'

AnatolianHieroglyphA481

\u{14618}: '𔘘'

AnatolianHieroglyphA482

\u{14619}: '𔘙'

AnatolianHieroglyphA483

\u{1461a}: '𔘚'

AnatolianHieroglyphA484

\u{1461b}: '𔘛'

AnatolianHieroglyphA485

\u{1461c}: '𔘜'

AnatolianHieroglyphA486

\u{1461d}: '𔘝'

AnatolianHieroglyphA487

\u{1461e}: '𔘞'

AnatolianHieroglyphA488

\u{1461f}: '𔘟'

AnatolianHieroglyphA489

\u{14620}: '𔘠'

AnatolianHieroglyphA490

\u{14621}: '𔘡'

AnatolianHieroglyphA491

\u{14622}: '𔘢'

AnatolianHieroglyphA492

\u{14623}: '𔘣'

AnatolianHieroglyphA493

\u{14624}: '𔘤'

AnatolianHieroglyphA494

\u{14625}: '𔘥'

AnatolianHieroglyphA495

\u{14626}: '𔘦'

AnatolianHieroglyphA496

\u{14627}: '𔘧'

AnatolianHieroglyphA497

\u{14628}: '𔘨'

AnatolianHieroglyphA501

\u{14629}: '𔘩'

AnatolianHieroglyphA502

\u{1462a}: '𔘪'

AnatolianHieroglyphA503

\u{1462b}: '𔘫'

AnatolianHieroglyphA504

\u{1462c}: '𔘬'

AnatolianHieroglyphA505

\u{1462d}: '𔘭'

AnatolianHieroglyphA506

\u{1462e}: '𔘮'

AnatolianHieroglyphA507

\u{1462f}: '𔘯'

AnatolianHieroglyphA508

\u{14630}: '𔘰'

AnatolianHieroglyphA509

\u{14631}: '𔘱'

AnatolianHieroglyphA510

\u{14632}: '𔘲'

AnatolianHieroglyphA511

\u{14633}: '𔘳'

AnatolianHieroglyphA512

\u{14634}: '𔘴'

AnatolianHieroglyphA513

\u{14635}: '𔘵'

AnatolianHieroglyphA514

\u{14636}: '𔘶'

AnatolianHieroglyphA515

\u{14637}: '𔘷'

AnatolianHieroglyphA516

\u{14638}: '𔘸'

AnatolianHieroglyphA517

\u{14639}: '𔘹'

AnatolianHieroglyphA518

\u{1463a}: '𔘺'

AnatolianHieroglyphA519

\u{1463b}: '𔘻'

AnatolianHieroglyphA520

\u{1463c}: '𔘼'

AnatolianHieroglyphA521

\u{1463d}: '𔘽'

AnatolianHieroglyphA522

\u{1463e}: '𔘾'

AnatolianHieroglyphA523

\u{1463f}: '𔘿'

AnatolianHieroglyphA524

\u{14640}: '𔙀'

AnatolianHieroglyphA525

\u{14641}: '𔙁'

AnatolianHieroglyphA526

\u{14642}: '𔙂'

AnatolianHieroglyphA527

\u{14643}: '𔙃'

AnatolianHieroglyphA528

\u{14644}: '𔙄'

AnatolianHieroglyphA529

\u{14645}: '𔙅'

AnatolianHieroglyphA530

\u{14646}: '𔙆'

Methods

impl AnatolianHieroglyphs

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case