Enum unicode_types::arabic::Arabic

pub enum Arabic {
    NumberSign,
    SignSanah,
    FootnoteMarker,
    SignSafha,
    SignSamvat,
    NumberMarkAbove,
    DashIndicCubeRoot,
    DashIndicFourthRoot,
    Ray,
    DashIndicPerMilleSign,
    DashIndicPerTenThousandSign,
    AfghaniSign,
    Comma,
    DateSeparator,
    PoeticVerseSign,
    SignMisra,
    SignSallallahouAlayheWassallam,
    SignAlayheAssallam,
    SignRahmatullahAlayhe,
    SignRadiAllahouAnhu,
    SignTakhallus,
    SmallHighTah,
    SmallHighLigatureAlefWithLamWithYeh,
    SmallHighZain,
    SmallFatha,
    SmallDamma,
    SmallKasra,
    Semicolon,
    LetterMark,
    TripleDotPunctuationMark,
    QuestionMark,
    LetterKashmiriYeh,
    LetterHamza,
    LetterAlefWithMaddaAbove,
    LetterAlefWithHamzaAbove,
    LetterWawWithHamzaAbove,
    LetterAlefWithHamzaBelow,
    LetterYehWithHamzaAbove,
    LetterAlef,
    LetterBeh,
    LetterTehMarbuta,
    LetterTeh,
    LetterTheh,
    LetterJeem,
    LetterHah,
    LetterKhah,
    LetterDal,
    LetterThal,
    LetterReh,
    LetterZain,
    LetterSeen,
    LetterSheen,
    LetterSad,
    LetterDad,
    LetterTah,
    LetterZah,
    LetterAin,
    LetterGhain,
    LetterKehehWithTwoDotsAbove,
    LetterKehehWithThreeDotsBelow,
    LetterFarsiYehWithInvertedV,
    LetterFarsiYehWithTwoDotsAbove,
    LetterFarsiYehWithThreeDotsAbove,
    Tatweel,
    LetterFeh,
    LetterQaf,
    LetterKaf,
    LetterLam,
    LetterMeem,
    LetterNoon,
    LetterHeh,
    LetterWaw,
    LetterAlefMaksura,
    LetterYeh,
    Fathatan,
    Dammatan,
    Kasratan,
    Fatha,
    Damma,
    Kasra,
    Shadda,
    Sukun,
    MaddahAbove,
    HamzaAbove,
    HamzaBelow,
    SubscriptAlef,
    InvertedDamma,
    MarkNoonGhunna,
    Zwarakay,
    VowelSignSmallVAbove,
    VowelSignInvertedSmallVAbove,
    VowelSignDotBelow,
    ReversedDamma,
    FathaWithTwoDots,
    WavyHamzaBelow,
    DashIndicDigitZero,
    DashIndicDigitOne,
    DashIndicDigitTwo,
    DashIndicDigitThree,
    DashIndicDigitFour,
    DashIndicDigitFive,
    DashIndicDigitSix,
    DashIndicDigitSeven,
    DashIndicDigitEight,
    DashIndicDigitNine,
    PercentSign,
    DecimalSeparator,
    ThousandsSeparator,
    FivePointedStar,
    LetterDotlessBeh,
    LetterDotlessQaf,
    LetterSuperscriptAlef,
    LetterAlefWasla,
    LetterAlefWithWavyHamzaAbove,
    LetterAlefWithWavyHamzaBelow,
    LetterHighHamza,
    LetterHighHamzaAlef,
    LetterHighHamzaWaw,
    LetterUWithHamzaAbove,
    LetterHighHamzaYeh,
    LetterTteh,
    LetterTteheh,
    LetterBeeh,
    LetterTehWithRing,
    LetterTehWithThreeDotsAboveDownwards,
    LetterPeh,
    LetterTeheh,
    LetterBeheh,
    LetterHahWithHamzaAbove,
    LetterHahWithTwoDotsVerticalAbove,
    LetterNyeh,
    LetterDyeh,
    LetterHahWithThreeDotsAbove,
    LetterTcheh,
    LetterTcheheh,
    LetterDdal,
    LetterDalWithRing,
    LetterDalWithDotBelow,
    LetterDalWithDotBelowAndSmallTah,
    LetterDahal,
    LetterDdahal,
    LetterDul,
    LetterDalWithThreeDotsAboveDownwards,
    LetterDalWithFourDotsAbove,
    LetterRreh,
    LetterRehWithSmallV,
    LetterRehWithRing,
    LetterRehWithDotBelow,
    LetterRehWithSmallVBelow,
    LetterRehWithDotBelowAndDotAbove,
    LetterRehWithTwoDotsAbove,
    LetterJeh,
    LetterRehWithFourDotsAbove,
    LetterSeenWithDotBelowAndDotAbove,
    LetterSeenWithThreeDotsBelow,
    LetterSeenWithThreeDotsBelowAndThreeDotsAbove,
    LetterSadWithTwoDotsBelow,
    LetterSadWithThreeDotsAbove,
    LetterTahWithThreeDotsAbove,
    LetterAinWithThreeDotsAbove,
    LetterDotlessFeh,
    LetterFehWithDotMovedBelow,
    LetterFehWithDotBelow,
    LetterVeh,
    LetterFehWithThreeDotsBelow,
    LetterPeheh,
    LetterQafWithDotAbove,
    LetterQafWithThreeDotsAbove,
    LetterKeheh,
    LetterSwashKaf,
    LetterKafWithRing,
    LetterKafWithDotAbove,
    LetterNg,
    LetterKafWithThreeDotsBelow,
    LetterGaf,
    LetterGafWithRing,
    LetterNgoeh,
    LetterGafWithTwoDotsBelow,
    LetterGueh,
    LetterGafWithThreeDotsAbove,
    LetterLamWithSmallV,
    LetterLamWithDotAbove,
    LetterLamWithThreeDotsAbove,
    LetterLamWithThreeDotsBelow,
    LetterNoonWithDotBelow,
    LetterNoonGhunna,
    LetterRnoon,
    LetterNoonWithRing,
    LetterNoonWithThreeDotsAbove,
    LetterHehDoachashmee,
    LetterTchehWithDotAbove,
    LetterHehWithYehAbove,
    LetterHehGoal,
    LetterHehGoalWithHamzaAbove,
    LetterTehMarbutaGoal,
    LetterWawWithRing,
    LetterKirghizOe,
    LetterOe,
    LetterU,
    LetterYu,
    LetterKirghizYu,
    LetterWawWithTwoDotsAbove,
    LetterVe,
    LetterFarsiYeh,
    LetterYehWithTail,
    LetterYehWithSmallV,
    LetterWawWithDotAbove,
    LetterE,
    LetterYehWithThreeDotsBelow,
    LetterYehBarree,
    LetterYehBarreeWithHamzaAbove,
    FullStop,
    LetterAe,
    SmallHighLigatureSadWithLamWithAlefMaksura,
    SmallHighLigatureQafWithLamWithAlefMaksura,
    SmallHighMeemInitialForm,
    SmallHighLamAlef,
    SmallHighJeem,
    SmallHighThreeDots,
    SmallHighSeen,
    EndOfAyah,
    StartOfRubElHizb,
    SmallHighRoundedZero,
    SmallHighUprightRectangularZero,
    SmallHighDotlessHeadOfKhah,
    SmallHighMeemIsolatedForm,
    SmallLowSeen,
    SmallHighMadda,
    SmallWaw,
    SmallYeh,
    SmallHighYeh,
    SmallHighNoon,
    PlaceOfSajdah,
    EmptyCentreLowStop,
    EmptyCentreHighStop,
    RoundedHighStopWithFilledCentre,
    SmallLowMeem,
    LetterDalWithInvertedV,
    LetterRehWithInvertedV,
    ExtendedDashIndicDigitZero,
    ExtendedDashIndicDigitOne,
    ExtendedDashIndicDigitTwo,
    ExtendedDashIndicDigitThree,
    ExtendedDashIndicDigitFour,
    ExtendedDashIndicDigitFive,
    ExtendedDashIndicDigitSix,
    ExtendedDashIndicDigitSeven,
    ExtendedDashIndicDigitEight,
    ExtendedDashIndicDigitNine,
    LetterSheenWithDotBelow,
    LetterDadWithDotBelow,
    LetterGhainWithDotBelow,
    SignSindhiAmpersand,
    SignSindhiPostpositionMen,
}

An enum to represent all characters in the Arabic block.

Variants
NumberSign

\u{600}: '؀'

SignSanah

\u{601}: '؁'

FootnoteMarker

\u{602}: '؂'

SignSafha

\u{603}: '؃'

SignSamvat

\u{604}: '؄'

NumberMarkAbove

\u{605}: '؅'

DashIndicCubeRoot

\u{606}: '؆'

DashIndicFourthRoot

\u{607}: '؇'

Ray

\u{608}: '؈'

DashIndicPerMilleSign

\u{609}: '؉'

DashIndicPerTenThousandSign

\u{60a}: '؊'

AfghaniSign

\u{60b}: '؋'

Comma

\u{60c}: '،'

DateSeparator

\u{60d}: '؍'

PoeticVerseSign

\u{60e}: '؎'

SignMisra

\u{60f}: '؏'

SignSallallahouAlayheWassallam

\u{610}: 'ؐ'

SignAlayheAssallam

\u{611}: 'ؑ'

SignRahmatullahAlayhe

\u{612}: 'ؒ'

SignRadiAllahouAnhu

\u{613}: 'ؓ'

SignTakhallus

\u{614}: 'ؔ'

SmallHighTah

\u{615}: 'ؕ'

SmallHighLigatureAlefWithLamWithYeh

\u{616}: 'ؖ'

SmallHighZain

\u{617}: 'ؗ'

SmallFatha

\u{618}: 'ؘ'

SmallDamma

\u{619}: 'ؙ'

SmallKasra

\u{61a}: 'ؚ'

Semicolon

\u{61b}: '؛'

LetterMark

\u{61c}: '؜'

TripleDotPunctuationMark

\u{61e}: '؞'

QuestionMark

\u{61f}: '؟'

LetterKashmiriYeh

\u{620}: 'ؠ'

LetterHamza

\u{621}: 'ء'

LetterAlefWithMaddaAbove

\u{622}: 'آ'

LetterAlefWithHamzaAbove

\u{623}: 'أ'

LetterWawWithHamzaAbove

\u{624}: 'ؤ'

LetterAlefWithHamzaBelow

\u{625}: 'إ'

LetterYehWithHamzaAbove

\u{626}: 'ئ'

LetterAlef

\u{627}: 'ا'

LetterBeh

\u{628}: 'ب'

LetterTehMarbuta

\u{629}: 'ة'

LetterTeh

\u{62a}: 'ت'

LetterTheh

\u{62b}: 'ث'

LetterJeem

\u{62c}: 'ج'

LetterHah

\u{62d}: 'ح'

LetterKhah

\u{62e}: 'خ'

LetterDal

\u{62f}: 'د'

LetterThal

\u{630}: 'ذ'

LetterReh

\u{631}: 'ر'

LetterZain

\u{632}: 'ز'

LetterSeen

\u{633}: 'س'

LetterSheen

\u{634}: 'ش'

LetterSad

\u{635}: 'ص'

LetterDad

\u{636}: 'ض'

LetterTah

\u{637}: 'ط'

LetterZah

\u{638}: 'ظ'

LetterAin

\u{639}: 'ع'

LetterGhain

\u{63a}: 'غ'

LetterKehehWithTwoDotsAbove

\u{63b}: 'ػ'

LetterKehehWithThreeDotsBelow

\u{63c}: 'ؼ'

LetterFarsiYehWithInvertedV

\u{63d}: 'ؽ'

LetterFarsiYehWithTwoDotsAbove

\u{63e}: 'ؾ'

LetterFarsiYehWithThreeDotsAbove

\u{63f}: 'ؿ'

Tatweel

\u{640}: 'ـ'

LetterFeh

\u{641}: 'ف'

LetterQaf

\u{642}: 'ق'

LetterKaf

\u{643}: 'ك'

LetterLam

\u{644}: 'ل'

LetterMeem

\u{645}: 'م'

LetterNoon

\u{646}: 'ن'

LetterHeh

\u{647}: 'ه'

LetterWaw

\u{648}: 'و'

LetterAlefMaksura

\u{649}: 'ى'

LetterYeh

\u{64a}: 'ي'

Fathatan

\u{64b}: 'ً'

Dammatan

\u{64c}: 'ٌ'

Kasratan

\u{64d}: 'ٍ'

Fatha

\u{64e}: 'َ'

Damma

\u{64f}: 'ُ'

Kasra

\u{650}: 'ِ'

Shadda

\u{651}: 'ّ'

Sukun

\u{652}: 'ْ'

MaddahAbove

\u{653}: 'ٓ'

HamzaAbove

\u{654}: 'ٔ'

HamzaBelow

\u{655}: 'ٕ'

SubscriptAlef

\u{656}: 'ٖ'

InvertedDamma

\u{657}: 'ٗ'

MarkNoonGhunna

\u{658}: '٘'

Zwarakay

\u{659}: 'ٙ'

VowelSignSmallVAbove

\u{65a}: 'ٚ'

VowelSignInvertedSmallVAbove

\u{65b}: 'ٛ'

VowelSignDotBelow

\u{65c}: 'ٜ'

ReversedDamma

\u{65d}: 'ٝ'

FathaWithTwoDots

\u{65e}: 'ٞ'

WavyHamzaBelow

\u{65f}: 'ٟ'

DashIndicDigitZero

\u{660}: '٠'

DashIndicDigitOne

\u{661}: '١'

DashIndicDigitTwo

\u{662}: '٢'

DashIndicDigitThree

\u{663}: '٣'

DashIndicDigitFour

\u{664}: '٤'

DashIndicDigitFive

\u{665}: '٥'

DashIndicDigitSix

\u{666}: '٦'

DashIndicDigitSeven

\u{667}: '٧'

DashIndicDigitEight

\u{668}: '٨'

DashIndicDigitNine

\u{669}: '٩'

PercentSign

\u{66a}: '٪'

DecimalSeparator

\u{66b}: '٫'

ThousandsSeparator

\u{66c}: '٬'

FivePointedStar

\u{66d}: '٭'

LetterDotlessBeh

\u{66e}: 'ٮ'

LetterDotlessQaf

\u{66f}: 'ٯ'

LetterSuperscriptAlef

\u{670}: 'ٰ'

LetterAlefWasla

\u{671}: 'ٱ'

LetterAlefWithWavyHamzaAbove

\u{672}: 'ٲ'

LetterAlefWithWavyHamzaBelow

\u{673}: 'ٳ'

LetterHighHamza

\u{674}: 'ٴ'

LetterHighHamzaAlef

\u{675}: 'ٵ'

LetterHighHamzaWaw

\u{676}: 'ٶ'

LetterUWithHamzaAbove

\u{677}: 'ٷ'

LetterHighHamzaYeh

\u{678}: 'ٸ'

LetterTteh

\u{679}: 'ٹ'

LetterTteheh

\u{67a}: 'ٺ'

LetterBeeh

\u{67b}: 'ٻ'

LetterTehWithRing

\u{67c}: 'ټ'

LetterTehWithThreeDotsAboveDownwards

\u{67d}: 'ٽ'

LetterPeh

\u{67e}: 'پ'

LetterTeheh

\u{67f}: 'ٿ'

LetterBeheh

\u{680}: 'ڀ'

LetterHahWithHamzaAbove

\u{681}: 'ځ'

LetterHahWithTwoDotsVerticalAbove

\u{682}: 'ڂ'

LetterNyeh

\u{683}: 'ڃ'

LetterDyeh

\u{684}: 'ڄ'

LetterHahWithThreeDotsAbove

\u{685}: 'څ'

LetterTcheh

\u{686}: 'چ'

LetterTcheheh

\u{687}: 'ڇ'

LetterDdal

\u{688}: 'ڈ'

LetterDalWithRing

\u{689}: 'ډ'

LetterDalWithDotBelow

\u{68a}: 'ڊ'

LetterDalWithDotBelowAndSmallTah

\u{68b}: 'ڋ'

LetterDahal

\u{68c}: 'ڌ'

LetterDdahal

\u{68d}: 'ڍ'

LetterDul

\u{68e}: 'ڎ'

LetterDalWithThreeDotsAboveDownwards

\u{68f}: 'ڏ'

LetterDalWithFourDotsAbove

\u{690}: 'ڐ'

LetterRreh

\u{691}: 'ڑ'

LetterRehWithSmallV

\u{692}: 'ڒ'

LetterRehWithRing

\u{693}: 'ړ'

LetterRehWithDotBelow

\u{694}: 'ڔ'

LetterRehWithSmallVBelow

\u{695}: 'ڕ'

LetterRehWithDotBelowAndDotAbove

\u{696}: 'ږ'

LetterRehWithTwoDotsAbove

\u{697}: 'ڗ'

LetterJeh

\u{698}: 'ژ'

LetterRehWithFourDotsAbove

\u{699}: 'ڙ'

LetterSeenWithDotBelowAndDotAbove

\u{69a}: 'ښ'

LetterSeenWithThreeDotsBelow

\u{69b}: 'ڛ'

LetterSeenWithThreeDotsBelowAndThreeDotsAbove

\u{69c}: 'ڜ'

LetterSadWithTwoDotsBelow

\u{69d}: 'ڝ'

LetterSadWithThreeDotsAbove

\u{69e}: 'ڞ'

LetterTahWithThreeDotsAbove

\u{69f}: 'ڟ'

LetterAinWithThreeDotsAbove

\u{6a0}: 'ڠ'

LetterDotlessFeh

\u{6a1}: 'ڡ'

LetterFehWithDotMovedBelow

\u{6a2}: 'ڢ'

LetterFehWithDotBelow

\u{6a3}: 'ڣ'

LetterVeh

\u{6a4}: 'ڤ'

LetterFehWithThreeDotsBelow

\u{6a5}: 'ڥ'

LetterPeheh

\u{6a6}: 'ڦ'

LetterQafWithDotAbove

\u{6a7}: 'ڧ'

LetterQafWithThreeDotsAbove

\u{6a8}: 'ڨ'

LetterKeheh

\u{6a9}: 'ک'

LetterSwashKaf

\u{6aa}: 'ڪ'

LetterKafWithRing

\u{6ab}: 'ګ'

LetterKafWithDotAbove

\u{6ac}: 'ڬ'

LetterNg

\u{6ad}: 'ڭ'

LetterKafWithThreeDotsBelow

\u{6ae}: 'ڮ'

LetterGaf

\u{6af}: 'گ'

LetterGafWithRing

\u{6b0}: 'ڰ'

LetterNgoeh

\u{6b1}: 'ڱ'

LetterGafWithTwoDotsBelow

\u{6b2}: 'ڲ'

LetterGueh

\u{6b3}: 'ڳ'

LetterGafWithThreeDotsAbove

\u{6b4}: 'ڴ'

LetterLamWithSmallV

\u{6b5}: 'ڵ'

LetterLamWithDotAbove

\u{6b6}: 'ڶ'

LetterLamWithThreeDotsAbove

\u{6b7}: 'ڷ'

LetterLamWithThreeDotsBelow

\u{6b8}: 'ڸ'

LetterNoonWithDotBelow

\u{6b9}: 'ڹ'

LetterNoonGhunna

\u{6ba}: 'ں'

LetterRnoon

\u{6bb}: 'ڻ'

LetterNoonWithRing

\u{6bc}: 'ڼ'

LetterNoonWithThreeDotsAbove

\u{6bd}: 'ڽ'

LetterHehDoachashmee

\u{6be}: 'ھ'

LetterTchehWithDotAbove

\u{6bf}: 'ڿ'

LetterHehWithYehAbove

\u{6c0}: 'ۀ'

LetterHehGoal

\u{6c1}: 'ہ'

LetterHehGoalWithHamzaAbove

\u{6c2}: 'ۂ'

LetterTehMarbutaGoal

\u{6c3}: 'ۃ'

LetterWawWithRing

\u{6c4}: 'ۄ'

LetterKirghizOe

\u{6c5}: 'ۅ'

LetterOe

\u{6c6}: 'ۆ'

LetterU

\u{6c7}: 'ۇ'

LetterYu

\u{6c8}: 'ۈ'

LetterKirghizYu

\u{6c9}: 'ۉ'

LetterWawWithTwoDotsAbove

\u{6ca}: 'ۊ'

LetterVe

\u{6cb}: 'ۋ'

LetterFarsiYeh

\u{6cc}: 'ی'

LetterYehWithTail

\u{6cd}: 'ۍ'

LetterYehWithSmallV

\u{6ce}: 'ێ'

LetterWawWithDotAbove

\u{6cf}: 'ۏ'

LetterE

\u{6d0}: 'ې'

LetterYehWithThreeDotsBelow

\u{6d1}: 'ۑ'

LetterYehBarree

\u{6d2}: 'ے'

LetterYehBarreeWithHamzaAbove

\u{6d3}: 'ۓ'

FullStop

\u{6d4}: '۔'

LetterAe

\u{6d5}: 'ە'

SmallHighLigatureSadWithLamWithAlefMaksura

\u{6d6}: 'ۖ'

SmallHighLigatureQafWithLamWithAlefMaksura

\u{6d7}: 'ۗ'

SmallHighMeemInitialForm

\u{6d8}: 'ۘ'

SmallHighLamAlef

\u{6d9}: 'ۙ'

SmallHighJeem

\u{6da}: 'ۚ'

SmallHighThreeDots

\u{6db}: 'ۛ'

SmallHighSeen

\u{6dc}: 'ۜ'

EndOfAyah

\u{6dd}: '۝'

StartOfRubElHizb

\u{6de}: '۞'

SmallHighRoundedZero

\u{6df}: '۟'

SmallHighUprightRectangularZero

\u{6e0}: '۠'

SmallHighDotlessHeadOfKhah

\u{6e1}: 'ۡ'

SmallHighMeemIsolatedForm

\u{6e2}: 'ۢ'

SmallLowSeen

\u{6e3}: 'ۣ'

SmallHighMadda

\u{6e4}: 'ۤ'

SmallWaw

\u{6e5}: 'ۥ'

SmallYeh

\u{6e6}: 'ۦ'

SmallHighYeh

\u{6e7}: 'ۧ'

SmallHighNoon

\u{6e8}: 'ۨ'

PlaceOfSajdah

\u{6e9}: '۩'

EmptyCentreLowStop

\u{6ea}: '۪'

EmptyCentreHighStop

\u{6eb}: '۫'

RoundedHighStopWithFilledCentre

\u{6ec}: '۬'

SmallLowMeem

\u{6ed}: 'ۭ'

LetterDalWithInvertedV

\u{6ee}: 'ۮ'

LetterRehWithInvertedV

\u{6ef}: 'ۯ'

ExtendedDashIndicDigitZero

\u{6f0}: '۰'

ExtendedDashIndicDigitOne

\u{6f1}: '۱'

ExtendedDashIndicDigitTwo

\u{6f2}: '۲'

ExtendedDashIndicDigitThree

\u{6f3}: '۳'

ExtendedDashIndicDigitFour

\u{6f4}: '۴'

ExtendedDashIndicDigitFive

\u{6f5}: '۵'

ExtendedDashIndicDigitSix

\u{6f6}: '۶'

ExtendedDashIndicDigitSeven

\u{6f7}: '۷'

ExtendedDashIndicDigitEight

\u{6f8}: '۸'

ExtendedDashIndicDigitNine

\u{6f9}: '۹'

LetterSheenWithDotBelow

\u{6fa}: 'ۺ'

LetterDadWithDotBelow

\u{6fb}: 'ۻ'

LetterGhainWithDotBelow

\u{6fc}: 'ۼ'

SignSindhiAmpersand

\u{6fd}: '۽'

SignSindhiPostpositionMen

\u{6fe}: '۾'

Methods

impl Arabic

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case