Enum unicode_types::arabic_mathematical_alphabetic_symbols::ArabicMathematicalAlphabeticSymbols

pub enum ArabicMathematicalAlphabeticSymbols {
    ArabicMathematicalAlef,
    ArabicMathematicalBeh,
    ArabicMathematicalJeem,
    ArabicMathematicalDal,
    ArabicMathematicalWaw,
    ArabicMathematicalZain,
    ArabicMathematicalHah,
    ArabicMathematicalTah,
    ArabicMathematicalYeh,
    ArabicMathematicalKaf,
    ArabicMathematicalLam,
    ArabicMathematicalMeem,
    ArabicMathematicalNoon,
    ArabicMathematicalSeen,
    ArabicMathematicalAin,
    ArabicMathematicalFeh,
    ArabicMathematicalSad,
    ArabicMathematicalQaf,
    ArabicMathematicalReh,
    ArabicMathematicalSheen,
    ArabicMathematicalTeh,
    ArabicMathematicalTheh,
    ArabicMathematicalKhah,
    ArabicMathematicalThal,
    ArabicMathematicalDad,
    ArabicMathematicalZah,
    ArabicMathematicalGhain,
    ArabicMathematicalDotlessBeh,
    ArabicMathematicalDotlessNoon,
    ArabicMathematicalDotlessFeh,
    ArabicMathematicalDotlessQaf,
    ArabicMathematicalInitialBeh,
    ArabicMathematicalInitialJeem,
    ArabicMathematicalInitialHeh,
    ArabicMathematicalInitialHah,
    ArabicMathematicalInitialYeh,
    ArabicMathematicalInitialKaf,
    ArabicMathematicalInitialLam,
    ArabicMathematicalInitialMeem,
    ArabicMathematicalInitialNoon,
    ArabicMathematicalInitialSeen,
    ArabicMathematicalInitialAin,
    ArabicMathematicalInitialFeh,
    ArabicMathematicalInitialSad,
    ArabicMathematicalInitialQaf,
    ArabicMathematicalInitialSheen,
    ArabicMathematicalInitialTeh,
    ArabicMathematicalInitialTheh,
    ArabicMathematicalInitialKhah,
    ArabicMathematicalInitialDad,
    ArabicMathematicalInitialGhain,
    ArabicMathematicalTailedJeem,
    ArabicMathematicalTailedHah,
    ArabicMathematicalTailedYeh,
    ArabicMathematicalTailedLam,
    ArabicMathematicalTailedNoon,
    ArabicMathematicalTailedSeen,
    ArabicMathematicalTailedAin,
    ArabicMathematicalTailedSad,
    ArabicMathematicalTailedQaf,
    ArabicMathematicalTailedSheen,
    ArabicMathematicalTailedKhah,
    ArabicMathematicalTailedDad,
    ArabicMathematicalTailedGhain,
    ArabicMathematicalTailedDotlessNoon,
    ArabicMathematicalTailedDotlessQaf,
    ArabicMathematicalStretchedBeh,
    ArabicMathematicalStretchedJeem,
    ArabicMathematicalStretchedHeh,
    ArabicMathematicalStretchedHah,
    ArabicMathematicalStretchedTah,
    ArabicMathematicalStretchedYeh,
    ArabicMathematicalStretchedKaf,
    ArabicMathematicalStretchedMeem,
    ArabicMathematicalStretchedNoon,
    ArabicMathematicalStretchedSeen,
    ArabicMathematicalStretchedAin,
    ArabicMathematicalStretchedFeh,
    ArabicMathematicalStretchedSad,
    ArabicMathematicalStretchedQaf,
    ArabicMathematicalStretchedSheen,
    ArabicMathematicalStretchedTeh,
    ArabicMathematicalStretchedTheh,
    ArabicMathematicalStretchedKhah,
    ArabicMathematicalStretchedDad,
    ArabicMathematicalStretchedZah,
    ArabicMathematicalStretchedGhain,
    ArabicMathematicalStretchedDotlessBeh,
    ArabicMathematicalStretchedDotlessFeh,
    ArabicMathematicalLoopedAlef,
    ArabicMathematicalLoopedBeh,
    ArabicMathematicalLoopedJeem,
    ArabicMathematicalLoopedDal,
    ArabicMathematicalLoopedHeh,
    ArabicMathematicalLoopedWaw,
    ArabicMathematicalLoopedZain,
    ArabicMathematicalLoopedHah,
    ArabicMathematicalLoopedTah,
    ArabicMathematicalLoopedYeh,
    ArabicMathematicalLoopedLam,
    ArabicMathematicalLoopedMeem,
    ArabicMathematicalLoopedNoon,
    ArabicMathematicalLoopedSeen,
    ArabicMathematicalLoopedAin,
    ArabicMathematicalLoopedFeh,
    ArabicMathematicalLoopedSad,
    ArabicMathematicalLoopedQaf,
    ArabicMathematicalLoopedReh,
    ArabicMathematicalLoopedSheen,
    ArabicMathematicalLoopedTeh,
    ArabicMathematicalLoopedTheh,
    ArabicMathematicalLoopedKhah,
    ArabicMathematicalLoopedThal,
    ArabicMathematicalLoopedDad,
    ArabicMathematicalLoopedZah,
    ArabicMathematicalLoopedGhain,
    ArabicMathematicalDoubleDashStruckBeh,
    ArabicMathematicalDoubleDashStruckJeem,
    ArabicMathematicalDoubleDashStruckDal,
    ArabicMathematicalDoubleDashStruckWaw,
    ArabicMathematicalDoubleDashStruckZain,
    ArabicMathematicalDoubleDashStruckHah,
    ArabicMathematicalDoubleDashStruckTah,
    ArabicMathematicalDoubleDashStruckYeh,
    ArabicMathematicalDoubleDashStruckLam,
    ArabicMathematicalDoubleDashStruckMeem,
    ArabicMathematicalDoubleDashStruckNoon,
    ArabicMathematicalDoubleDashStruckSeen,
    ArabicMathematicalDoubleDashStruckAin,
    ArabicMathematicalDoubleDashStruckFeh,
    ArabicMathematicalDoubleDashStruckSad,
    ArabicMathematicalDoubleDashStruckQaf,
    ArabicMathematicalDoubleDashStruckReh,
    ArabicMathematicalDoubleDashStruckSheen,
    ArabicMathematicalDoubleDashStruckTeh,
    ArabicMathematicalDoubleDashStruckTheh,
    ArabicMathematicalDoubleDashStruckKhah,
    ArabicMathematicalDoubleDashStruckThal,
    ArabicMathematicalDoubleDashStruckDad,
    ArabicMathematicalDoubleDashStruckZah,
    ArabicMathematicalDoubleDashStruckGhain,
    ArabicMathematicalOperatorMeemWithHahWithTatweel,
    ArabicMathematicalOperatorHahWithDal,
}

An enum to represent all characters in the ArabicMathematicalAlphabeticSymbols block.

Variants
ArabicMathematicalAlef

\u{1ee00}: '𞸀'

ArabicMathematicalBeh

\u{1ee01}: '𞸁'

ArabicMathematicalJeem

\u{1ee02}: '𞸂'

ArabicMathematicalDal

\u{1ee03}: '𞸃'

ArabicMathematicalWaw

\u{1ee05}: '𞸅'

ArabicMathematicalZain

\u{1ee06}: '𞸆'

ArabicMathematicalHah

\u{1ee07}: '𞸇'

ArabicMathematicalTah

\u{1ee08}: '𞸈'

ArabicMathematicalYeh

\u{1ee09}: '𞸉'

ArabicMathematicalKaf

\u{1ee0a}: '𞸊'

ArabicMathematicalLam

\u{1ee0b}: '𞸋'

ArabicMathematicalMeem

\u{1ee0c}: '𞸌'

ArabicMathematicalNoon

\u{1ee0d}: '𞸍'

ArabicMathematicalSeen

\u{1ee0e}: '𞸎'

ArabicMathematicalAin

\u{1ee0f}: '𞸏'

ArabicMathematicalFeh

\u{1ee10}: '𞸐'

ArabicMathematicalSad

\u{1ee11}: '𞸑'

ArabicMathematicalQaf

\u{1ee12}: '𞸒'

ArabicMathematicalReh

\u{1ee13}: '𞸓'

ArabicMathematicalSheen

\u{1ee14}: '𞸔'

ArabicMathematicalTeh

\u{1ee15}: '𞸕'

ArabicMathematicalTheh

\u{1ee16}: '𞸖'

ArabicMathematicalKhah

\u{1ee17}: '𞸗'

ArabicMathematicalThal

\u{1ee18}: '𞸘'

ArabicMathematicalDad

\u{1ee19}: '𞸙'

ArabicMathematicalZah

\u{1ee1a}: '𞸚'

ArabicMathematicalGhain

\u{1ee1b}: '𞸛'

ArabicMathematicalDotlessBeh

\u{1ee1c}: '𞸜'

ArabicMathematicalDotlessNoon

\u{1ee1d}: '𞸝'

ArabicMathematicalDotlessFeh

\u{1ee1e}: '𞸞'

ArabicMathematicalDotlessQaf

\u{1ee1f}: '𞸟'

ArabicMathematicalInitialBeh

\u{1ee21}: '𞸡'

ArabicMathematicalInitialJeem

\u{1ee22}: '𞸢'

ArabicMathematicalInitialHeh

\u{1ee24}: '𞸤'

ArabicMathematicalInitialHah

\u{1ee27}: '𞸧'

ArabicMathematicalInitialYeh

\u{1ee29}: '𞸩'

ArabicMathematicalInitialKaf

\u{1ee2a}: '𞸪'

ArabicMathematicalInitialLam

\u{1ee2b}: '𞸫'

ArabicMathematicalInitialMeem

\u{1ee2c}: '𞸬'

ArabicMathematicalInitialNoon

\u{1ee2d}: '𞸭'

ArabicMathematicalInitialSeen

\u{1ee2e}: '𞸮'

ArabicMathematicalInitialAin

\u{1ee2f}: '𞸯'

ArabicMathematicalInitialFeh

\u{1ee30}: '𞸰'

ArabicMathematicalInitialSad

\u{1ee31}: '𞸱'

ArabicMathematicalInitialQaf

\u{1ee32}: '𞸲'

ArabicMathematicalInitialSheen

\u{1ee34}: '𞸴'

ArabicMathematicalInitialTeh

\u{1ee35}: '𞸵'

ArabicMathematicalInitialTheh

\u{1ee36}: '𞸶'

ArabicMathematicalInitialKhah

\u{1ee37}: '𞸷'

ArabicMathematicalInitialDad

\u{1ee39}: '𞸹'

ArabicMathematicalInitialGhain

\u{1ee3b}: '𞸻'

ArabicMathematicalTailedJeem

\u{1ee42}: '𞹂'

ArabicMathematicalTailedHah

\u{1ee47}: '𞹇'

ArabicMathematicalTailedYeh

\u{1ee49}: '𞹉'

ArabicMathematicalTailedLam

\u{1ee4b}: '𞹋'

ArabicMathematicalTailedNoon

\u{1ee4d}: '𞹍'

ArabicMathematicalTailedSeen

\u{1ee4e}: '𞹎'

ArabicMathematicalTailedAin

\u{1ee4f}: '𞹏'

ArabicMathematicalTailedSad

\u{1ee51}: '𞹑'

ArabicMathematicalTailedQaf

\u{1ee52}: '𞹒'

ArabicMathematicalTailedSheen

\u{1ee54}: '𞹔'

ArabicMathematicalTailedKhah

\u{1ee57}: '𞹗'

ArabicMathematicalTailedDad

\u{1ee59}: '𞹙'

ArabicMathematicalTailedGhain

\u{1ee5b}: '𞹛'

ArabicMathematicalTailedDotlessNoon

\u{1ee5d}: '𞹝'

ArabicMathematicalTailedDotlessQaf

\u{1ee5f}: '𞹟'

ArabicMathematicalStretchedBeh

\u{1ee61}: '𞹡'

ArabicMathematicalStretchedJeem

\u{1ee62}: '𞹢'

ArabicMathematicalStretchedHeh

\u{1ee64}: '𞹤'

ArabicMathematicalStretchedHah

\u{1ee67}: '𞹧'

ArabicMathematicalStretchedTah

\u{1ee68}: '𞹨'

ArabicMathematicalStretchedYeh

\u{1ee69}: '𞹩'

ArabicMathematicalStretchedKaf

\u{1ee6a}: '𞹪'

ArabicMathematicalStretchedMeem

\u{1ee6c}: '𞹬'

ArabicMathematicalStretchedNoon

\u{1ee6d}: '𞹭'

ArabicMathematicalStretchedSeen

\u{1ee6e}: '𞹮'

ArabicMathematicalStretchedAin

\u{1ee6f}: '𞹯'

ArabicMathematicalStretchedFeh

\u{1ee70}: '𞹰'

ArabicMathematicalStretchedSad

\u{1ee71}: '𞹱'

ArabicMathematicalStretchedQaf

\u{1ee72}: '𞹲'

ArabicMathematicalStretchedSheen

\u{1ee74}: '𞹴'

ArabicMathematicalStretchedTeh

\u{1ee75}: '𞹵'

ArabicMathematicalStretchedTheh

\u{1ee76}: '𞹶'

ArabicMathematicalStretchedKhah

\u{1ee77}: '𞹷'

ArabicMathematicalStretchedDad

\u{1ee79}: '𞹹'

ArabicMathematicalStretchedZah

\u{1ee7a}: '𞹺'

ArabicMathematicalStretchedGhain

\u{1ee7b}: '𞹻'

ArabicMathematicalStretchedDotlessBeh

\u{1ee7c}: '𞹼'

ArabicMathematicalStretchedDotlessFeh

\u{1ee7e}: '𞹾'

ArabicMathematicalLoopedAlef

\u{1ee80}: '𞺀'

ArabicMathematicalLoopedBeh

\u{1ee81}: '𞺁'

ArabicMathematicalLoopedJeem

\u{1ee82}: '𞺂'

ArabicMathematicalLoopedDal

\u{1ee83}: '𞺃'

ArabicMathematicalLoopedHeh

\u{1ee84}: '𞺄'

ArabicMathematicalLoopedWaw

\u{1ee85}: '𞺅'

ArabicMathematicalLoopedZain

\u{1ee86}: '𞺆'

ArabicMathematicalLoopedHah

\u{1ee87}: '𞺇'

ArabicMathematicalLoopedTah

\u{1ee88}: '𞺈'

ArabicMathematicalLoopedYeh

\u{1ee89}: '𞺉'

ArabicMathematicalLoopedLam

\u{1ee8b}: '𞺋'

ArabicMathematicalLoopedMeem

\u{1ee8c}: '𞺌'

ArabicMathematicalLoopedNoon

\u{1ee8d}: '𞺍'

ArabicMathematicalLoopedSeen

\u{1ee8e}: '𞺎'

ArabicMathematicalLoopedAin

\u{1ee8f}: '𞺏'

ArabicMathematicalLoopedFeh

\u{1ee90}: '𞺐'

ArabicMathematicalLoopedSad

\u{1ee91}: '𞺑'

ArabicMathematicalLoopedQaf

\u{1ee92}: '𞺒'

ArabicMathematicalLoopedReh

\u{1ee93}: '𞺓'

ArabicMathematicalLoopedSheen

\u{1ee94}: '𞺔'

ArabicMathematicalLoopedTeh

\u{1ee95}: '𞺕'

ArabicMathematicalLoopedTheh

\u{1ee96}: '𞺖'

ArabicMathematicalLoopedKhah

\u{1ee97}: '𞺗'

ArabicMathematicalLoopedThal

\u{1ee98}: '𞺘'

ArabicMathematicalLoopedDad

\u{1ee99}: '𞺙'

ArabicMathematicalLoopedZah

\u{1ee9a}: '𞺚'

ArabicMathematicalLoopedGhain

\u{1ee9b}: '𞺛'

ArabicMathematicalDoubleDashStruckBeh

\u{1eea1}: '𞺡'

ArabicMathematicalDoubleDashStruckJeem

\u{1eea2}: '𞺢'

ArabicMathematicalDoubleDashStruckDal

\u{1eea3}: '𞺣'

ArabicMathematicalDoubleDashStruckWaw

\u{1eea5}: '𞺥'

ArabicMathematicalDoubleDashStruckZain

\u{1eea6}: '𞺦'

ArabicMathematicalDoubleDashStruckHah

\u{1eea7}: '𞺧'

ArabicMathematicalDoubleDashStruckTah

\u{1eea8}: '𞺨'

ArabicMathematicalDoubleDashStruckYeh

\u{1eea9}: '𞺩'

ArabicMathematicalDoubleDashStruckLam

\u{1eeab}: '𞺫'

ArabicMathematicalDoubleDashStruckMeem

\u{1eeac}: '𞺬'

ArabicMathematicalDoubleDashStruckNoon

\u{1eead}: '𞺭'

ArabicMathematicalDoubleDashStruckSeen

\u{1eeae}: '𞺮'

ArabicMathematicalDoubleDashStruckAin

\u{1eeaf}: '𞺯'

ArabicMathematicalDoubleDashStruckFeh

\u{1eeb0}: '𞺰'

ArabicMathematicalDoubleDashStruckSad

\u{1eeb1}: '𞺱'

ArabicMathematicalDoubleDashStruckQaf

\u{1eeb2}: '𞺲'

ArabicMathematicalDoubleDashStruckReh

\u{1eeb3}: '𞺳'

ArabicMathematicalDoubleDashStruckSheen

\u{1eeb4}: '𞺴'

ArabicMathematicalDoubleDashStruckTeh

\u{1eeb5}: '𞺵'

ArabicMathematicalDoubleDashStruckTheh

\u{1eeb6}: '𞺶'

ArabicMathematicalDoubleDashStruckKhah

\u{1eeb7}: '𞺷'

ArabicMathematicalDoubleDashStruckThal

\u{1eeb8}: '𞺸'

ArabicMathematicalDoubleDashStruckDad

\u{1eeb9}: '𞺹'

ArabicMathematicalDoubleDashStruckZah

\u{1eeba}: '𞺺'

ArabicMathematicalDoubleDashStruckGhain

\u{1eebb}: '𞺻'

ArabicMathematicalOperatorMeemWithHahWithTatweel

\u{1eef0}: '𞻰'

ArabicMathematicalOperatorHahWithDal

\u{1eef1}: '𞻱'

Methods

impl ArabicMathematicalAlphabeticSymbols

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case