Enum unicode_types::arabic_presentation_forms_a::ArabicPresentationFormsA

pub enum ArabicPresentationFormsA {
    ArabicLetterAlefWaslaIsolatedForm,
    ArabicLetterAlefWaslaFinalForm,
    ArabicLetterBeehIsolatedForm,
    ArabicLetterBeehFinalForm,
    ArabicLetterBeehInitialForm,
    ArabicLetterBeehMedialForm,
    ArabicLetterPehIsolatedForm,
    ArabicLetterPehFinalForm,
    ArabicLetterPehInitialForm,
    ArabicLetterPehMedialForm,
    ArabicLetterBehehIsolatedForm,
    ArabicLetterBehehFinalForm,
    ArabicLetterBehehInitialForm,
    ArabicLetterBehehMedialForm,
    ArabicLetterTtehehIsolatedForm,
    ArabicLetterTtehehFinalForm,
    ArabicLetterTtehehInitialForm,
    ArabicLetterTtehehMedialForm,
    ArabicLetterTehehIsolatedForm,
    ArabicLetterTehehFinalForm,
    ArabicLetterTehehInitialForm,
    ArabicLetterTehehMedialForm,
    ArabicLetterTtehIsolatedForm,
    ArabicLetterTtehFinalForm,
    ArabicLetterTtehInitialForm,
    ArabicLetterTtehMedialForm,
    ArabicLetterVehIsolatedForm,
    ArabicLetterVehFinalForm,
    ArabicLetterVehInitialForm,
    ArabicLetterVehMedialForm,
    ArabicLetterPehehIsolatedForm,
    ArabicLetterPehehFinalForm,
    ArabicLetterPehehInitialForm,
    ArabicLetterPehehMedialForm,
    ArabicLetterDyehIsolatedForm,
    ArabicLetterDyehFinalForm,
    ArabicLetterDyehInitialForm,
    ArabicLetterDyehMedialForm,
    ArabicLetterNyehIsolatedForm,
    ArabicLetterNyehFinalForm,
    ArabicLetterNyehInitialForm,
    ArabicLetterNyehMedialForm,
    ArabicLetterTchehIsolatedForm,
    ArabicLetterTchehFinalForm,
    ArabicLetterTchehInitialForm,
    ArabicLetterTchehMedialForm,
    ArabicLetterTchehehIsolatedForm,
    ArabicLetterTchehehFinalForm,
    ArabicLetterTchehehInitialForm,
    ArabicLetterTchehehMedialForm,
    ArabicLetterDdahalIsolatedForm,
    ArabicLetterDdahalFinalForm,
    ArabicLetterDahalIsolatedForm,
    ArabicLetterDahalFinalForm,
    ArabicLetterDulIsolatedForm,
    ArabicLetterDulFinalForm,
    ArabicLetterDdalIsolatedForm,
    ArabicLetterDdalFinalForm,
    ArabicLetterJehIsolatedForm,
    ArabicLetterJehFinalForm,
    ArabicLetterRrehIsolatedForm,
    ArabicLetterRrehFinalForm,
    ArabicLetterKehehIsolatedForm,
    ArabicLetterKehehFinalForm,
    ArabicLetterKehehInitialForm,
    ArabicLetterKehehMedialForm,
    ArabicLetterGafIsolatedForm,
    ArabicLetterGafFinalForm,
    ArabicLetterGafInitialForm,
    ArabicLetterGafMedialForm,
    ArabicLetterGuehIsolatedForm,
    ArabicLetterGuehFinalForm,
    ArabicLetterGuehInitialForm,
    ArabicLetterGuehMedialForm,
    ArabicLetterNgoehIsolatedForm,
    ArabicLetterNgoehFinalForm,
    ArabicLetterNgoehInitialForm,
    ArabicLetterNgoehMedialForm,
    ArabicLetterNoonGhunnaIsolatedForm,
    ArabicLetterNoonGhunnaFinalForm,
    ArabicLetterRnoonIsolatedForm,
    ArabicLetterRnoonFinalForm,
    ArabicLetterRnoonInitialForm,
    ArabicLetterRnoonMedialForm,
    ArabicLetterHehWithYehAboveIsolatedForm,
    ArabicLetterHehWithYehAboveFinalForm,
    ArabicLetterHehGoalIsolatedForm,
    ArabicLetterHehGoalFinalForm,
    ArabicLetterHehGoalInitialForm,
    ArabicLetterHehGoalMedialForm,
    ArabicLetterHehDoachashmeeIsolatedForm,
    ArabicLetterHehDoachashmeeFinalForm,
    ArabicLetterHehDoachashmeeInitialForm,
    ArabicLetterHehDoachashmeeMedialForm,
    ArabicLetterYehBarreeIsolatedForm,
    ArabicLetterYehBarreeFinalForm,
    ArabicLetterYehBarreeWithHamzaAboveIsolatedForm,
    ArabicLetterYehBarreeWithHamzaAboveFinalForm,
    ArabicSymbolDotAbove,
    ArabicSymbolDotBelow,
    ArabicSymbolTwoDotsAbove,
    ArabicSymbolTwoDotsBelow,
    ArabicSymbolThreeDotsAbove,
    ArabicSymbolThreeDotsBelow,
    ArabicSymbolThreeDotsPointingDownwardsAbove,
    ArabicSymbolThreeDotsPointingDownwardsBelow,
    ArabicSymbolFourDotsAbove,
    ArabicSymbolFourDotsBelow,
    ArabicSymbolDoubleVerticalBarBelow,
    ArabicSymbolTwoDotsVerticallyAbove,
    ArabicSymbolTwoDotsVerticallyBelow,
    ArabicSymbolRing,
    ArabicSymbolSmallTahAbove,
    ArabicSymbolSmallTahBelow,
    ArabicLetterNgIsolatedForm,
    ArabicLetterNgFinalForm,
    ArabicLetterNgInitialForm,
    ArabicLetterNgMedialForm,
    ArabicLetterUIsolatedForm,
    ArabicLetterUFinalForm,
    ArabicLetterOeIsolatedForm,
    ArabicLetterOeFinalForm,
    ArabicLetterYuIsolatedForm,
    ArabicLetterYuFinalForm,
    ArabicLetterUWithHamzaAboveIsolatedForm,
    ArabicLetterVeIsolatedForm,
    ArabicLetterVeFinalForm,
    ArabicLetterKirghizOeIsolatedForm,
    ArabicLetterKirghizOeFinalForm,
    ArabicLetterKirghizYuIsolatedForm,
    ArabicLetterKirghizYuFinalForm,
    ArabicLetterEIsolatedForm,
    ArabicLetterEFinalForm,
    ArabicLetterEInitialForm,
    ArabicLetterEMedialForm,
    ArabicLetterUighurKazakhKirghizAlefMaksuraInitialForm,
    ArabicLetterUighurKazakhKirghizAlefMaksuraMedialForm,
    ArabicLigatureYehWithHamzaAboveWithAlefIsolatedForm,
    ArabicLigatureYehWithHamzaAboveWithAlefFinalForm,
    ArabicLigatureYehWithHamzaAboveWithAeIsolatedForm,
    ArabicLigatureYehWithHamzaAboveWithAeFinalForm,
    ArabicLigatureYehWithHamzaAboveWithWawIsolatedForm,
    ArabicLigatureYehWithHamzaAboveWithWawFinalForm,
    ArabicLigatureYehWithHamzaAboveWithUIsolatedForm,
    ArabicLigatureYehWithHamzaAboveWithUFinalForm,
    ArabicLigatureYehWithHamzaAboveWithOeIsolatedForm,
    ArabicLigatureYehWithHamzaAboveWithOeFinalForm,
    ArabicLigatureYehWithHamzaAboveWithYuIsolatedForm,
    ArabicLigatureYehWithHamzaAboveWithYuFinalForm,
    ArabicLigatureYehWithHamzaAboveWithEIsolatedForm,
    ArabicLigatureYehWithHamzaAboveWithEFinalForm,
    ArabicLigatureYehWithHamzaAboveWithEInitialForm,
    ArabicLigatureUighurKirghizYehWithHamzaAboveWithAlefMaksuraIsolatedForm,
    ArabicLigatureUighurKirghizYehWithHamzaAboveWithAlefMaksuraFinalForm,
    ArabicLigatureUighurKirghizYehWithHamzaAboveWithAlefMaksuraInitialForm,
    ArabicLetterFarsiYehIsolatedForm,
    ArabicLetterFarsiYehFinalForm,
    ArabicLetterFarsiYehInitialForm,
    ArabicLetterFarsiYehMedialForm,
    ArabicLigatureYehWithHamzaAboveWithJeemIsolatedForm,
    ArabicLigatureYehWithHamzaAboveWithHahIsolatedForm,
    ArabicLigatureYehWithHamzaAboveWithMeemIsolatedForm,
    ArabicLigatureYehWithHamzaAboveWithAlefMaksuraIsolatedForm,
    ArabicLigatureYehWithHamzaAboveWithYehIsolatedForm,
    ArabicLigatureBehWithJeemIsolatedForm,
    ArabicLigatureBehWithHahIsolatedForm,
    ArabicLigatureBehWithKhahIsolatedForm,
    ArabicLigatureBehWithMeemIsolatedForm,
    ArabicLigatureBehWithAlefMaksuraIsolatedForm,
    ArabicLigatureBehWithYehIsolatedForm,
    ArabicLigatureTehWithJeemIsolatedForm,
    ArabicLigatureTehWithHahIsolatedForm,
    ArabicLigatureTehWithKhahIsolatedForm,
    ArabicLigatureTehWithMeemIsolatedForm,
    ArabicLigatureTehWithAlefMaksuraIsolatedForm,
    ArabicLigatureTehWithYehIsolatedForm,
    ArabicLigatureThehWithJeemIsolatedForm,
    ArabicLigatureThehWithMeemIsolatedForm,
    ArabicLigatureThehWithAlefMaksuraIsolatedForm,
    ArabicLigatureThehWithYehIsolatedForm,
    ArabicLigatureJeemWithHahIsolatedForm,
    ArabicLigatureJeemWithMeemIsolatedForm,
    ArabicLigatureHahWithJeemIsolatedForm,
    ArabicLigatureHahWithMeemIsolatedForm,
    ArabicLigatureKhahWithJeemIsolatedForm,
    ArabicLigatureKhahWithHahIsolatedForm,
    ArabicLigatureKhahWithMeemIsolatedForm,
    ArabicLigatureSeenWithJeemIsolatedForm,
    ArabicLigatureSeenWithHahIsolatedForm,
    ArabicLigatureSeenWithKhahIsolatedForm,
    ArabicLigatureSeenWithMeemIsolatedForm,
    ArabicLigatureSadWithHahIsolatedForm,
    ArabicLigatureSadWithMeemIsolatedForm,
    ArabicLigatureDadWithJeemIsolatedForm,
    ArabicLigatureDadWithHahIsolatedForm,
    ArabicLigatureDadWithKhahIsolatedForm,
    ArabicLigatureDadWithMeemIsolatedForm,
    ArabicLigatureTahWithHahIsolatedForm,
    ArabicLigatureTahWithMeemIsolatedForm,
    ArabicLigatureZahWithMeemIsolatedForm,
    ArabicLigatureAinWithJeemIsolatedForm,
    ArabicLigatureAinWithMeemIsolatedForm,
    ArabicLigatureGhainWithJeemIsolatedForm,
    ArabicLigatureGhainWithMeemIsolatedForm,
    ArabicLigatureFehWithJeemIsolatedForm,
    ArabicLigatureFehWithHahIsolatedForm,
    ArabicLigatureFehWithKhahIsolatedForm,
    ArabicLigatureFehWithMeemIsolatedForm,
    ArabicLigatureFehWithAlefMaksuraIsolatedForm,
    ArabicLigatureFehWithYehIsolatedForm,
    ArabicLigatureQafWithHahIsolatedForm,
    ArabicLigatureQafWithMeemIsolatedForm,
    ArabicLigatureQafWithAlefMaksuraIsolatedForm,
    ArabicLigatureQafWithYehIsolatedForm,
    ArabicLigatureKafWithAlefIsolatedForm,
    ArabicLigatureKafWithJeemIsolatedForm,
    ArabicLigatureKafWithHahIsolatedForm,
    ArabicLigatureKafWithKhahIsolatedForm,
    ArabicLigatureKafWithLamIsolatedForm,
    ArabicLigatureKafWithMeemIsolatedForm,
    ArabicLigatureKafWithAlefMaksuraIsolatedForm,
    ArabicLigatureKafWithYehIsolatedForm,
    ArabicLigatureLamWithJeemIsolatedForm,
    ArabicLigatureLamWithHahIsolatedForm,
    ArabicLigatureLamWithKhahIsolatedForm,
    ArabicLigatureLamWithMeemIsolatedForm,
    ArabicLigatureLamWithAlefMaksuraIsolatedForm,
    ArabicLigatureLamWithYehIsolatedForm,
    ArabicLigatureMeemWithJeemIsolatedForm,
    ArabicLigatureMeemWithHahIsolatedForm,
    ArabicLigatureMeemWithKhahIsolatedForm,
    ArabicLigatureMeemWithMeemIsolatedForm,
    ArabicLigatureMeemWithAlefMaksuraIsolatedForm,
    ArabicLigatureMeemWithYehIsolatedForm,
    ArabicLigatureNoonWithJeemIsolatedForm,
    ArabicLigatureNoonWithHahIsolatedForm,
    ArabicLigatureNoonWithKhahIsolatedForm,
    ArabicLigatureNoonWithMeemIsolatedForm,
    ArabicLigatureNoonWithAlefMaksuraIsolatedForm,
    ArabicLigatureNoonWithYehIsolatedForm,
    ArabicLigatureHehWithJeemIsolatedForm,
    ArabicLigatureHehWithMeemIsolatedForm,
    ArabicLigatureHehWithAlefMaksuraIsolatedForm,
    ArabicLigatureHehWithYehIsolatedForm,
    ArabicLigatureYehWithJeemIsolatedForm,
    ArabicLigatureYehWithHahIsolatedForm,
    ArabicLigatureYehWithKhahIsolatedForm,
    ArabicLigatureYehWithMeemIsolatedForm,
    ArabicLigatureYehWithAlefMaksuraIsolatedForm,
    ArabicLigatureYehWithYehIsolatedForm,
    ArabicLigatureThalWithSuperscriptAlefIsolatedForm,
    ArabicLigatureRehWithSuperscriptAlefIsolatedForm,
    ArabicLigatureAlefMaksuraWithSuperscriptAlefIsolatedForm,
    ArabicLigatureShaddaWithDammatanIsolatedForm,
    ArabicLigatureShaddaWithKasratanIsolatedForm,
    ArabicLigatureShaddaWithFathaIsolatedForm,
    ArabicLigatureShaddaWithDammaIsolatedForm,
    ArabicLigatureShaddaWithKasraIsolatedForm,
    ArabicLigatureShaddaWithSuperscriptAlefIsolatedForm,
    ArabicLigatureYehWithHamzaAboveWithRehFinalForm,
    ArabicLigatureYehWithHamzaAboveWithZainFinalForm,
    ArabicLigatureYehWithHamzaAboveWithMeemFinalForm,
    ArabicLigatureYehWithHamzaAboveWithNoonFinalForm,
    ArabicLigatureYehWithHamzaAboveWithAlefMaksuraFinalForm,
    ArabicLigatureYehWithHamzaAboveWithYehFinalForm,
    ArabicLigatureBehWithRehFinalForm,
    ArabicLigatureBehWithZainFinalForm,
    ArabicLigatureBehWithMeemFinalForm,
    ArabicLigatureBehWithNoonFinalForm,
    ArabicLigatureBehWithAlefMaksuraFinalForm,
    ArabicLigatureBehWithYehFinalForm,
    ArabicLigatureTehWithRehFinalForm,
    ArabicLigatureTehWithZainFinalForm,
    ArabicLigatureTehWithMeemFinalForm,
    ArabicLigatureTehWithNoonFinalForm,
    ArabicLigatureTehWithAlefMaksuraFinalForm,
    ArabicLigatureTehWithYehFinalForm,
    ArabicLigatureThehWithRehFinalForm,
    ArabicLigatureThehWithZainFinalForm,
    ArabicLigatureThehWithMeemFinalForm,
    ArabicLigatureThehWithNoonFinalForm,
    ArabicLigatureThehWithAlefMaksuraFinalForm,
    ArabicLigatureThehWithYehFinalForm,
    ArabicLigatureFehWithAlefMaksuraFinalForm,
    ArabicLigatureFehWithYehFinalForm,
    ArabicLigatureQafWithAlefMaksuraFinalForm,
    ArabicLigatureQafWithYehFinalForm,
    ArabicLigatureKafWithAlefFinalForm,
    ArabicLigatureKafWithLamFinalForm,
    ArabicLigatureKafWithMeemFinalForm,
    ArabicLigatureKafWithAlefMaksuraFinalForm,
    ArabicLigatureKafWithYehFinalForm,
    ArabicLigatureLamWithMeemFinalForm,
    ArabicLigatureLamWithAlefMaksuraFinalForm,
    ArabicLigatureLamWithYehFinalForm,
    ArabicLigatureMeemWithAlefFinalForm,
    ArabicLigatureMeemWithMeemFinalForm,
    ArabicLigatureNoonWithRehFinalForm,
    ArabicLigatureNoonWithZainFinalForm,
    ArabicLigatureNoonWithMeemFinalForm,
    ArabicLigatureNoonWithNoonFinalForm,
    ArabicLigatureNoonWithAlefMaksuraFinalForm,
    ArabicLigatureNoonWithYehFinalForm,
    ArabicLigatureAlefMaksuraWithSuperscriptAlefFinalForm,
    ArabicLigatureYehWithRehFinalForm,
    ArabicLigatureYehWithZainFinalForm,
    ArabicLigatureYehWithMeemFinalForm,
    ArabicLigatureYehWithNoonFinalForm,
    ArabicLigatureYehWithAlefMaksuraFinalForm,
    ArabicLigatureYehWithYehFinalForm,
    ArabicLigatureYehWithHamzaAboveWithJeemInitialForm,
    ArabicLigatureYehWithHamzaAboveWithHahInitialForm,
    ArabicLigatureYehWithHamzaAboveWithKhahInitialForm,
    ArabicLigatureYehWithHamzaAboveWithMeemInitialForm,
    ArabicLigatureYehWithHamzaAboveWithHehInitialForm,
    ArabicLigatureBehWithJeemInitialForm,
    ArabicLigatureBehWithHahInitialForm,
    ArabicLigatureBehWithKhahInitialForm,
    ArabicLigatureBehWithMeemInitialForm,
    ArabicLigatureBehWithHehInitialForm,
    ArabicLigatureTehWithJeemInitialForm,
    ArabicLigatureTehWithHahInitialForm,
    ArabicLigatureTehWithKhahInitialForm,
    ArabicLigatureTehWithMeemInitialForm,
    ArabicLigatureTehWithHehInitialForm,
    ArabicLigatureThehWithMeemInitialForm,
    ArabicLigatureJeemWithHahInitialForm,
    ArabicLigatureJeemWithMeemInitialForm,
    ArabicLigatureHahWithJeemInitialForm,
    ArabicLigatureHahWithMeemInitialForm,
    ArabicLigatureKhahWithJeemInitialForm,
    ArabicLigatureKhahWithMeemInitialForm,
    ArabicLigatureSeenWithJeemInitialForm,
    ArabicLigatureSeenWithHahInitialForm,
    ArabicLigatureSeenWithKhahInitialForm,
    ArabicLigatureSeenWithMeemInitialForm,
    ArabicLigatureSadWithHahInitialForm,
    ArabicLigatureSadWithKhahInitialForm,
    ArabicLigatureSadWithMeemInitialForm,
    ArabicLigatureDadWithJeemInitialForm,
    ArabicLigatureDadWithHahInitialForm,
    ArabicLigatureDadWithKhahInitialForm,
    ArabicLigatureDadWithMeemInitialForm,
    ArabicLigatureTahWithHahInitialForm,
    ArabicLigatureZahWithMeemInitialForm,
    ArabicLigatureAinWithJeemInitialForm,
    ArabicLigatureAinWithMeemInitialForm,
    ArabicLigatureGhainWithJeemInitialForm,
    ArabicLigatureGhainWithMeemInitialForm,
    ArabicLigatureFehWithJeemInitialForm,
    ArabicLigatureFehWithHahInitialForm,
    ArabicLigatureFehWithKhahInitialForm,
    ArabicLigatureFehWithMeemInitialForm,
    ArabicLigatureQafWithHahInitialForm,
    ArabicLigatureQafWithMeemInitialForm,
    ArabicLigatureKafWithJeemInitialForm,
    ArabicLigatureKafWithHahInitialForm,
    ArabicLigatureKafWithKhahInitialForm,
    ArabicLigatureKafWithLamInitialForm,
    ArabicLigatureKafWithMeemInitialForm,
    ArabicLigatureLamWithJeemInitialForm,
    ArabicLigatureLamWithHahInitialForm,
    ArabicLigatureLamWithKhahInitialForm,
    ArabicLigatureLamWithMeemInitialForm,
    ArabicLigatureLamWithHehInitialForm,
    ArabicLigatureMeemWithJeemInitialForm,
    ArabicLigatureMeemWithHahInitialForm,
    ArabicLigatureMeemWithKhahInitialForm,
    ArabicLigatureMeemWithMeemInitialForm,
    ArabicLigatureNoonWithJeemInitialForm,
    ArabicLigatureNoonWithHahInitialForm,
    ArabicLigatureNoonWithKhahInitialForm,
    ArabicLigatureNoonWithMeemInitialForm,
    ArabicLigatureNoonWithHehInitialForm,
    ArabicLigatureHehWithJeemInitialForm,
    ArabicLigatureHehWithMeemInitialForm,
    ArabicLigatureHehWithSuperscriptAlefInitialForm,
    ArabicLigatureYehWithJeemInitialForm,
    ArabicLigatureYehWithHahInitialForm,
    ArabicLigatureYehWithKhahInitialForm,
    ArabicLigatureYehWithMeemInitialForm,
    ArabicLigatureYehWithHehInitialForm,
    ArabicLigatureYehWithHamzaAboveWithMeemMedialForm,
    ArabicLigatureYehWithHamzaAboveWithHehMedialForm,
    ArabicLigatureBehWithMeemMedialForm,
    ArabicLigatureBehWithHehMedialForm,
    ArabicLigatureTehWithMeemMedialForm,
    ArabicLigatureTehWithHehMedialForm,
    ArabicLigatureThehWithMeemMedialForm,
    ArabicLigatureThehWithHehMedialForm,
    ArabicLigatureSeenWithMeemMedialForm,
    ArabicLigatureSeenWithHehMedialForm,
    ArabicLigatureSheenWithMeemMedialForm,
    ArabicLigatureSheenWithHehMedialForm,
    ArabicLigatureKafWithLamMedialForm,
    ArabicLigatureKafWithMeemMedialForm,
    ArabicLigatureLamWithMeemMedialForm,
    ArabicLigatureNoonWithMeemMedialForm,
    ArabicLigatureNoonWithHehMedialForm,
    ArabicLigatureYehWithMeemMedialForm,
    ArabicLigatureYehWithHehMedialForm,
    ArabicLigatureShaddaWithFathaMedialForm,
    ArabicLigatureShaddaWithDammaMedialForm,
    ArabicLigatureShaddaWithKasraMedialForm,
    ArabicLigatureTahWithAlefMaksuraIsolatedForm,
    ArabicLigatureTahWithYehIsolatedForm,
    ArabicLigatureAinWithAlefMaksuraIsolatedForm,
    ArabicLigatureAinWithYehIsolatedForm,
    ArabicLigatureGhainWithAlefMaksuraIsolatedForm,
    ArabicLigatureGhainWithYehIsolatedForm,
    ArabicLigatureSeenWithAlefMaksuraIsolatedForm,
    ArabicLigatureSeenWithYehIsolatedForm,
    ArabicLigatureSheenWithAlefMaksuraIsolatedForm,
    ArabicLigatureSheenWithYehIsolatedForm,
    ArabicLigatureHahWithAlefMaksuraIsolatedForm,
    ArabicLigatureHahWithYehIsolatedForm,
    ArabicLigatureJeemWithAlefMaksuraIsolatedForm,
    ArabicLigatureJeemWithYehIsolatedForm,
    ArabicLigatureKhahWithAlefMaksuraIsolatedForm,
    ArabicLigatureKhahWithYehIsolatedForm,
    ArabicLigatureSadWithAlefMaksuraIsolatedForm,
    ArabicLigatureSadWithYehIsolatedForm,
    ArabicLigatureDadWithAlefMaksuraIsolatedForm,
    ArabicLigatureDadWithYehIsolatedForm,
    ArabicLigatureSheenWithJeemIsolatedForm,
    ArabicLigatureSheenWithHahIsolatedForm,
    ArabicLigatureSheenWithKhahIsolatedForm,
    ArabicLigatureSheenWithMeemIsolatedForm,
    ArabicLigatureSheenWithRehIsolatedForm,
    ArabicLigatureSeenWithRehIsolatedForm,
    ArabicLigatureSadWithRehIsolatedForm,
    ArabicLigatureDadWithRehIsolatedForm,
    ArabicLigatureTahWithAlefMaksuraFinalForm,
    ArabicLigatureTahWithYehFinalForm,
    ArabicLigatureAinWithAlefMaksuraFinalForm,
    ArabicLigatureAinWithYehFinalForm,
    ArabicLigatureGhainWithAlefMaksuraFinalForm,
    ArabicLigatureGhainWithYehFinalForm,
    ArabicLigatureSeenWithAlefMaksuraFinalForm,
    ArabicLigatureSeenWithYehFinalForm,
    ArabicLigatureSheenWithAlefMaksuraFinalForm,
    ArabicLigatureSheenWithYehFinalForm,
    ArabicLigatureHahWithAlefMaksuraFinalForm,
    ArabicLigatureHahWithYehFinalForm,
    ArabicLigatureJeemWithAlefMaksuraFinalForm,
    ArabicLigatureJeemWithYehFinalForm,
    ArabicLigatureKhahWithAlefMaksuraFinalForm,
    ArabicLigatureKhahWithYehFinalForm,
    ArabicLigatureSadWithAlefMaksuraFinalForm,
    ArabicLigatureSadWithYehFinalForm,
    ArabicLigatureDadWithAlefMaksuraFinalForm,
    ArabicLigatureDadWithYehFinalForm,
    ArabicLigatureSheenWithJeemFinalForm,
    ArabicLigatureSheenWithHahFinalForm,
    ArabicLigatureSheenWithKhahFinalForm,
    ArabicLigatureSheenWithMeemFinalForm,
    ArabicLigatureSheenWithRehFinalForm,
    ArabicLigatureSeenWithRehFinalForm,
    ArabicLigatureSadWithRehFinalForm,
    ArabicLigatureDadWithRehFinalForm,
    ArabicLigatureSheenWithJeemInitialForm,
    ArabicLigatureSheenWithHahInitialForm,
    ArabicLigatureSheenWithKhahInitialForm,
    ArabicLigatureSheenWithMeemInitialForm,
    ArabicLigatureSeenWithHehInitialForm,
    ArabicLigatureSheenWithHehInitialForm,
    ArabicLigatureTahWithMeemInitialForm,
    ArabicLigatureSeenWithJeemMedialForm,
    ArabicLigatureSeenWithHahMedialForm,
    ArabicLigatureSeenWithKhahMedialForm,
    ArabicLigatureSheenWithJeemMedialForm,
    ArabicLigatureSheenWithHahMedialForm,
    ArabicLigatureSheenWithKhahMedialForm,
    ArabicLigatureTahWithMeemMedialForm,
    ArabicLigatureZahWithMeemMedialForm,
    ArabicLigatureAlefWithFathatanFinalForm,
    ArabicLigatureAlefWithFathatanIsolatedForm,
    OrnateLeftParenthesis,
    OrnateRightParenthesis,
    ArabicLigatureTehWithJeemWithMeemInitialForm,
    ArabicLigatureTehWithHahWithJeemFinalForm,
    ArabicLigatureTehWithHahWithJeemInitialForm,
    ArabicLigatureTehWithHahWithMeemInitialForm,
    ArabicLigatureTehWithKhahWithMeemInitialForm,
    ArabicLigatureTehWithMeemWithJeemInitialForm,
    ArabicLigatureTehWithMeemWithHahInitialForm,
    ArabicLigatureTehWithMeemWithKhahInitialForm,
    ArabicLigatureJeemWithMeemWithHahFinalForm,
    ArabicLigatureJeemWithMeemWithHahInitialForm,
    ArabicLigatureHahWithMeemWithYehFinalForm,
    ArabicLigatureHahWithMeemWithAlefMaksuraFinalForm,
    ArabicLigatureSeenWithHahWithJeemInitialForm,
    ArabicLigatureSeenWithJeemWithHahInitialForm,
    ArabicLigatureSeenWithJeemWithAlefMaksuraFinalForm,
    ArabicLigatureSeenWithMeemWithHahFinalForm,
    ArabicLigatureSeenWithMeemWithHahInitialForm,
    ArabicLigatureSeenWithMeemWithJeemInitialForm,
    ArabicLigatureSeenWithMeemWithMeemFinalForm,
    ArabicLigatureSeenWithMeemWithMeemInitialForm,
    ArabicLigatureSadWithHahWithHahFinalForm,
    ArabicLigatureSadWithHahWithHahInitialForm,
    ArabicLigatureSadWithMeemWithMeemFinalForm,
    ArabicLigatureSheenWithHahWithMeemFinalForm,
    ArabicLigatureSheenWithHahWithMeemInitialForm,
    ArabicLigatureSheenWithJeemWithYehFinalForm,
    ArabicLigatureSheenWithMeemWithKhahFinalForm,
    ArabicLigatureSheenWithMeemWithKhahInitialForm,
    ArabicLigatureSheenWithMeemWithMeemFinalForm,
    ArabicLigatureSheenWithMeemWithMeemInitialForm,
    ArabicLigatureDadWithHahWithAlefMaksuraFinalForm,
    ArabicLigatureDadWithKhahWithMeemFinalForm,
    ArabicLigatureDadWithKhahWithMeemInitialForm,
    ArabicLigatureTahWithMeemWithHahFinalForm,
    ArabicLigatureTahWithMeemWithHahInitialForm,
    ArabicLigatureTahWithMeemWithMeemInitialForm,
    ArabicLigatureTahWithMeemWithYehFinalForm,
    ArabicLigatureAinWithJeemWithMeemFinalForm,
    ArabicLigatureAinWithMeemWithMeemFinalForm,
    ArabicLigatureAinWithMeemWithMeemInitialForm,
    ArabicLigatureAinWithMeemWithAlefMaksuraFinalForm,
    ArabicLigatureGhainWithMeemWithMeemFinalForm,
    ArabicLigatureGhainWithMeemWithYehFinalForm,
    ArabicLigatureGhainWithMeemWithAlefMaksuraFinalForm,
    ArabicLigatureFehWithKhahWithMeemFinalForm,
    ArabicLigatureFehWithKhahWithMeemInitialForm,
    ArabicLigatureQafWithMeemWithHahFinalForm,
    ArabicLigatureQafWithMeemWithMeemFinalForm,
    ArabicLigatureLamWithHahWithMeemFinalForm,
    ArabicLigatureLamWithHahWithYehFinalForm,
    ArabicLigatureLamWithHahWithAlefMaksuraFinalForm,
    ArabicLigatureLamWithJeemWithJeemInitialForm,
    ArabicLigatureLamWithJeemWithJeemFinalForm,
    ArabicLigatureLamWithKhahWithMeemFinalForm,
    ArabicLigatureLamWithKhahWithMeemInitialForm,
    ArabicLigatureLamWithMeemWithHahFinalForm,
    ArabicLigatureLamWithMeemWithHahInitialForm,
    ArabicLigatureMeemWithHahWithJeemInitialForm,
    ArabicLigatureMeemWithHahWithMeemInitialForm,
    ArabicLigatureMeemWithHahWithYehFinalForm,
    ArabicLigatureMeemWithJeemWithHahInitialForm,
    ArabicLigatureMeemWithJeemWithMeemInitialForm,
    ArabicLigatureMeemWithKhahWithJeemInitialForm,
    ArabicLigatureMeemWithKhahWithMeemInitialForm,
    ArabicLigatureMeemWithJeemWithKhahInitialForm,
    ArabicLigatureHehWithMeemWithJeemInitialForm,
    ArabicLigatureHehWithMeemWithMeemInitialForm,
    ArabicLigatureNoonWithHahWithMeemInitialForm,
    ArabicLigatureNoonWithHahWithAlefMaksuraFinalForm,
    ArabicLigatureNoonWithJeemWithMeemFinalForm,
    ArabicLigatureNoonWithJeemWithMeemInitialForm,
    ArabicLigatureNoonWithJeemWithAlefMaksuraFinalForm,
    ArabicLigatureNoonWithMeemWithYehFinalForm,
    ArabicLigatureNoonWithMeemWithAlefMaksuraFinalForm,
    ArabicLigatureYehWithMeemWithMeemFinalForm,
    ArabicLigatureYehWithMeemWithMeemInitialForm,
    ArabicLigatureBehWithKhahWithYehFinalForm,
    ArabicLigatureTehWithJeemWithYehFinalForm,
    ArabicLigatureTehWithJeemWithAlefMaksuraFinalForm,
    ArabicLigatureTehWithKhahWithYehFinalForm,
    ArabicLigatureTehWithKhahWithAlefMaksuraFinalForm,
    ArabicLigatureTehWithMeemWithYehFinalForm,
    ArabicLigatureTehWithMeemWithAlefMaksuraFinalForm,
    ArabicLigatureJeemWithMeemWithYehFinalForm,
    ArabicLigatureJeemWithHahWithAlefMaksuraFinalForm,
    ArabicLigatureJeemWithMeemWithAlefMaksuraFinalForm,
    ArabicLigatureSeenWithKhahWithAlefMaksuraFinalForm,
    ArabicLigatureSadWithHahWithYehFinalForm,
    ArabicLigatureSheenWithHahWithYehFinalForm,
    ArabicLigatureDadWithHahWithYehFinalForm,
    ArabicLigatureLamWithJeemWithYehFinalForm,
    ArabicLigatureLamWithMeemWithYehFinalForm,
    ArabicLigatureYehWithHahWithYehFinalForm,
    ArabicLigatureYehWithJeemWithYehFinalForm,
    ArabicLigatureYehWithMeemWithYehFinalForm,
    ArabicLigatureMeemWithMeemWithYehFinalForm,
    ArabicLigatureQafWithMeemWithYehFinalForm,
    ArabicLigatureNoonWithHahWithYehFinalForm,
    ArabicLigatureQafWithMeemWithHahInitialForm,
    ArabicLigatureLamWithHahWithMeemInitialForm,
    ArabicLigatureAinWithMeemWithYehFinalForm,
    ArabicLigatureKafWithMeemWithYehFinalForm,
    ArabicLigatureNoonWithJeemWithHahInitialForm,
    ArabicLigatureMeemWithKhahWithYehFinalForm,
    ArabicLigatureLamWithJeemWithMeemInitialForm,
    ArabicLigatureKafWithMeemWithMeemFinalForm,
    ArabicLigatureLamWithJeemWithMeemFinalForm,
    ArabicLigatureNoonWithJeemWithHahFinalForm,
    ArabicLigatureJeemWithHahWithYehFinalForm,
    ArabicLigatureHahWithJeemWithYehFinalForm,
    ArabicLigatureMeemWithJeemWithYehFinalForm,
    ArabicLigatureFehWithMeemWithYehFinalForm,
    ArabicLigatureBehWithHahWithYehFinalForm,
    ArabicLigatureKafWithMeemWithMeemInitialForm,
    ArabicLigatureAinWithJeemWithMeemInitialForm,
    ArabicLigatureSadWithMeemWithMeemInitialForm,
    ArabicLigatureSeenWithKhahWithYehFinalForm,
    ArabicLigatureNoonWithJeemWithYehFinalForm,
    ArabicLigatureSallaUsedAsKoranicStopSignIsolatedForm,
    ArabicLigatureQalaUsedAsKoranicStopSignIsolatedForm,
    ArabicLigatureAllahIsolatedForm,
    ArabicLigatureAkbarIsolatedForm,
    ArabicLigatureMohammadIsolatedForm,
    ArabicLigatureSalamIsolatedForm,
    ArabicLigatureRasoulIsolatedForm,
    ArabicLigatureAlayheIsolatedForm,
    ArabicLigatureWasallamIsolatedForm,
    ArabicLigatureSallaIsolatedForm,
    ArabicLigatureSallallahouAlayheWasallam,
    ArabicLigatureJallajalalouhou,
    RialSign,
    ArabicLigatureBismillahArDashRahmanArDashRaheem,
}

An enum to represent all characters in the ArabicPresentationFormsA block.

Variants
ArabicLetterAlefWaslaIsolatedForm

\u{fb50}: ''

ArabicLetterAlefWaslaFinalForm

\u{fb51}: ''

ArabicLetterBeehIsolatedForm

\u{fb52}: ''

ArabicLetterBeehFinalForm

\u{fb53}: ''

ArabicLetterBeehInitialForm

\u{fb54}: ''

ArabicLetterBeehMedialForm

\u{fb55}: ''

ArabicLetterPehIsolatedForm

\u{fb56}: ''

ArabicLetterPehFinalForm

\u{fb57}: ''

ArabicLetterPehInitialForm

\u{fb58}: ''

ArabicLetterPehMedialForm

\u{fb59}: ''

ArabicLetterBehehIsolatedForm

\u{fb5a}: ''

ArabicLetterBehehFinalForm

\u{fb5b}: ''

ArabicLetterBehehInitialForm

\u{fb5c}: ''

ArabicLetterBehehMedialForm

\u{fb5d}: ''

ArabicLetterTtehehIsolatedForm

\u{fb5e}: ''

ArabicLetterTtehehFinalForm

\u{fb5f}: ''

ArabicLetterTtehehInitialForm

\u{fb60}: ''

ArabicLetterTtehehMedialForm

\u{fb61}: ''

ArabicLetterTehehIsolatedForm

\u{fb62}: ''

ArabicLetterTehehFinalForm

\u{fb63}: ''

ArabicLetterTehehInitialForm

\u{fb64}: ''

ArabicLetterTehehMedialForm

\u{fb65}: ''

ArabicLetterTtehIsolatedForm

\u{fb66}: ''

ArabicLetterTtehFinalForm

\u{fb67}: ''

ArabicLetterTtehInitialForm

\u{fb68}: ''

ArabicLetterTtehMedialForm

\u{fb69}: ''

ArabicLetterVehIsolatedForm

\u{fb6a}: ''

ArabicLetterVehFinalForm

\u{fb6b}: ''

ArabicLetterVehInitialForm

\u{fb6c}: ''

ArabicLetterVehMedialForm

\u{fb6d}: ''

ArabicLetterPehehIsolatedForm

\u{fb6e}: ''

ArabicLetterPehehFinalForm

\u{fb6f}: ''

ArabicLetterPehehInitialForm

\u{fb70}: ''

ArabicLetterPehehMedialForm

\u{fb71}: ''

ArabicLetterDyehIsolatedForm

\u{fb72}: ''

ArabicLetterDyehFinalForm

\u{fb73}: ''

ArabicLetterDyehInitialForm

\u{fb74}: ''

ArabicLetterDyehMedialForm

\u{fb75}: ''

ArabicLetterNyehIsolatedForm

\u{fb76}: ''

ArabicLetterNyehFinalForm

\u{fb77}: ''

ArabicLetterNyehInitialForm

\u{fb78}: ''

ArabicLetterNyehMedialForm

\u{fb79}: ''

ArabicLetterTchehIsolatedForm

\u{fb7a}: ''

ArabicLetterTchehFinalForm

\u{fb7b}: ''

ArabicLetterTchehInitialForm

\u{fb7c}: ''

ArabicLetterTchehMedialForm

\u{fb7d}: ''

ArabicLetterTchehehIsolatedForm

\u{fb7e}: ''

ArabicLetterTchehehFinalForm

\u{fb7f}: ''

ArabicLetterTchehehInitialForm

\u{fb80}: ''

ArabicLetterTchehehMedialForm

\u{fb81}: ''

ArabicLetterDdahalIsolatedForm

\u{fb82}: ''

ArabicLetterDdahalFinalForm

\u{fb83}: ''

ArabicLetterDahalIsolatedForm

\u{fb84}: ''

ArabicLetterDahalFinalForm

\u{fb85}: ''

ArabicLetterDulIsolatedForm

\u{fb86}: ''

ArabicLetterDulFinalForm

\u{fb87}: ''

ArabicLetterDdalIsolatedForm

\u{fb88}: ''

ArabicLetterDdalFinalForm

\u{fb89}: ''

ArabicLetterJehIsolatedForm

\u{fb8a}: ''

ArabicLetterJehFinalForm

\u{fb8b}: ''

ArabicLetterRrehIsolatedForm

\u{fb8c}: ''

ArabicLetterRrehFinalForm

\u{fb8d}: ''

ArabicLetterKehehIsolatedForm

\u{fb8e}: ''

ArabicLetterKehehFinalForm

\u{fb8f}: ''

ArabicLetterKehehInitialForm

\u{fb90}: ''

ArabicLetterKehehMedialForm

\u{fb91}: ''

ArabicLetterGafIsolatedForm

\u{fb92}: ''

ArabicLetterGafFinalForm

\u{fb93}: ''

ArabicLetterGafInitialForm

\u{fb94}: ''

ArabicLetterGafMedialForm

\u{fb95}: ''

ArabicLetterGuehIsolatedForm

\u{fb96}: ''

ArabicLetterGuehFinalForm

\u{fb97}: ''

ArabicLetterGuehInitialForm

\u{fb98}: ''

ArabicLetterGuehMedialForm

\u{fb99}: ''

ArabicLetterNgoehIsolatedForm

\u{fb9a}: ''

ArabicLetterNgoehFinalForm

\u{fb9b}: ''

ArabicLetterNgoehInitialForm

\u{fb9c}: ''

ArabicLetterNgoehMedialForm

\u{fb9d}: ''

ArabicLetterNoonGhunnaIsolatedForm

\u{fb9e}: ''

ArabicLetterNoonGhunnaFinalForm

\u{fb9f}: ''

ArabicLetterRnoonIsolatedForm

\u{fba0}: ''

ArabicLetterRnoonFinalForm

\u{fba1}: ''

ArabicLetterRnoonInitialForm

\u{fba2}: ''

ArabicLetterRnoonMedialForm

\u{fba3}: ''

ArabicLetterHehWithYehAboveIsolatedForm

\u{fba4}: ''

ArabicLetterHehWithYehAboveFinalForm

\u{fba5}: ''

ArabicLetterHehGoalIsolatedForm

\u{fba6}: ''

ArabicLetterHehGoalFinalForm

\u{fba7}: ''

ArabicLetterHehGoalInitialForm

\u{fba8}: ''

ArabicLetterHehGoalMedialForm

\u{fba9}: ''

ArabicLetterHehDoachashmeeIsolatedForm

\u{fbaa}: ''

ArabicLetterHehDoachashmeeFinalForm

\u{fbab}: ''

ArabicLetterHehDoachashmeeInitialForm

\u{fbac}: ''

ArabicLetterHehDoachashmeeMedialForm

\u{fbad}: ''

ArabicLetterYehBarreeIsolatedForm

\u{fbae}: ''

ArabicLetterYehBarreeFinalForm

\u{fbaf}: ''

ArabicLetterYehBarreeWithHamzaAboveIsolatedForm

\u{fbb0}: ''

ArabicLetterYehBarreeWithHamzaAboveFinalForm

\u{fbb1}: ''

ArabicSymbolDotAbove

\u{fbb2}: ''

ArabicSymbolDotBelow

\u{fbb3}: ''

ArabicSymbolTwoDotsAbove

\u{fbb4}: ''

ArabicSymbolTwoDotsBelow

\u{fbb5}: ''

ArabicSymbolThreeDotsAbove

\u{fbb6}: ''

ArabicSymbolThreeDotsBelow

\u{fbb7}: ''

ArabicSymbolThreeDotsPointingDownwardsAbove

\u{fbb8}: ''

ArabicSymbolThreeDotsPointingDownwardsBelow

\u{fbb9}: ''

ArabicSymbolFourDotsAbove

\u{fbba}: ''

ArabicSymbolFourDotsBelow

\u{fbbb}: ''

ArabicSymbolDoubleVerticalBarBelow

\u{fbbc}: ''

ArabicSymbolTwoDotsVerticallyAbove

\u{fbbd}: ''

ArabicSymbolTwoDotsVerticallyBelow

\u{fbbe}: ''

ArabicSymbolRing

\u{fbbf}: ''

ArabicSymbolSmallTahAbove

\u{fbc0}: ''

ArabicSymbolSmallTahBelow

\u{fbc1}: ''

ArabicLetterNgIsolatedForm

\u{fbd3}: ''

ArabicLetterNgFinalForm

\u{fbd4}: ''

ArabicLetterNgInitialForm

\u{fbd5}: ''

ArabicLetterNgMedialForm

\u{fbd6}: ''

ArabicLetterUIsolatedForm

\u{fbd7}: ''

ArabicLetterUFinalForm

\u{fbd8}: ''

ArabicLetterOeIsolatedForm

\u{fbd9}: ''

ArabicLetterOeFinalForm

\u{fbda}: ''

ArabicLetterYuIsolatedForm

\u{fbdb}: ''

ArabicLetterYuFinalForm

\u{fbdc}: ''

ArabicLetterUWithHamzaAboveIsolatedForm

\u{fbdd}: ''

ArabicLetterVeIsolatedForm

\u{fbde}: ''

ArabicLetterVeFinalForm

\u{fbdf}: ''

ArabicLetterKirghizOeIsolatedForm

\u{fbe0}: ''

ArabicLetterKirghizOeFinalForm

\u{fbe1}: ''

ArabicLetterKirghizYuIsolatedForm

\u{fbe2}: ''

ArabicLetterKirghizYuFinalForm

\u{fbe3}: ''

ArabicLetterEIsolatedForm

\u{fbe4}: ''

ArabicLetterEFinalForm

\u{fbe5}: ''

ArabicLetterEInitialForm

\u{fbe6}: ''

ArabicLetterEMedialForm

\u{fbe7}: ''

ArabicLetterUighurKazakhKirghizAlefMaksuraInitialForm

\u{fbe8}: ''

ArabicLetterUighurKazakhKirghizAlefMaksuraMedialForm

\u{fbe9}: ''

ArabicLigatureYehWithHamzaAboveWithAlefIsolatedForm

\u{fbea}: ''

ArabicLigatureYehWithHamzaAboveWithAlefFinalForm

\u{fbeb}: ''

ArabicLigatureYehWithHamzaAboveWithAeIsolatedForm

\u{fbec}: ''

ArabicLigatureYehWithHamzaAboveWithAeFinalForm

\u{fbed}: ''

ArabicLigatureYehWithHamzaAboveWithWawIsolatedForm

\u{fbee}: ''

ArabicLigatureYehWithHamzaAboveWithWawFinalForm

\u{fbef}: ''

ArabicLigatureYehWithHamzaAboveWithUIsolatedForm

\u{fbf0}: ''

ArabicLigatureYehWithHamzaAboveWithUFinalForm

\u{fbf1}: ''

ArabicLigatureYehWithHamzaAboveWithOeIsolatedForm

\u{fbf2}: ''

ArabicLigatureYehWithHamzaAboveWithOeFinalForm

\u{fbf3}: ''

ArabicLigatureYehWithHamzaAboveWithYuIsolatedForm

\u{fbf4}: ''

ArabicLigatureYehWithHamzaAboveWithYuFinalForm

\u{fbf5}: ''

ArabicLigatureYehWithHamzaAboveWithEIsolatedForm

\u{fbf6}: ''

ArabicLigatureYehWithHamzaAboveWithEFinalForm

\u{fbf7}: ''

ArabicLigatureYehWithHamzaAboveWithEInitialForm

\u{fbf8}: ''

ArabicLigatureUighurKirghizYehWithHamzaAboveWithAlefMaksuraIsolatedForm

\u{fbf9}: ''

ArabicLigatureUighurKirghizYehWithHamzaAboveWithAlefMaksuraFinalForm

\u{fbfa}: ''

ArabicLigatureUighurKirghizYehWithHamzaAboveWithAlefMaksuraInitialForm

\u{fbfb}: ''

ArabicLetterFarsiYehIsolatedForm

\u{fbfc}: ''

ArabicLetterFarsiYehFinalForm

\u{fbfd}: ''

ArabicLetterFarsiYehInitialForm

\u{fbfe}: ''

ArabicLetterFarsiYehMedialForm

\u{fbff}: ''

ArabicLigatureYehWithHamzaAboveWithJeemIsolatedForm

\u{fc00}: ''

ArabicLigatureYehWithHamzaAboveWithHahIsolatedForm

\u{fc01}: ''

ArabicLigatureYehWithHamzaAboveWithMeemIsolatedForm

\u{fc02}: ''

ArabicLigatureYehWithHamzaAboveWithAlefMaksuraIsolatedForm

\u{fc03}: ''

ArabicLigatureYehWithHamzaAboveWithYehIsolatedForm

\u{fc04}: ''

ArabicLigatureBehWithJeemIsolatedForm

\u{fc05}: ''

ArabicLigatureBehWithHahIsolatedForm

\u{fc06}: ''

ArabicLigatureBehWithKhahIsolatedForm

\u{fc07}: ''

ArabicLigatureBehWithMeemIsolatedForm

\u{fc08}: ''

ArabicLigatureBehWithAlefMaksuraIsolatedForm

\u{fc09}: ''

ArabicLigatureBehWithYehIsolatedForm

\u{fc0a}: ''

ArabicLigatureTehWithJeemIsolatedForm

\u{fc0b}: ''

ArabicLigatureTehWithHahIsolatedForm

\u{fc0c}: ''

ArabicLigatureTehWithKhahIsolatedForm

\u{fc0d}: ''

ArabicLigatureTehWithMeemIsolatedForm

\u{fc0e}: ''

ArabicLigatureTehWithAlefMaksuraIsolatedForm

\u{fc0f}: ''

ArabicLigatureTehWithYehIsolatedForm

\u{fc10}: ''

ArabicLigatureThehWithJeemIsolatedForm

\u{fc11}: ''

ArabicLigatureThehWithMeemIsolatedForm

\u{fc12}: ''

ArabicLigatureThehWithAlefMaksuraIsolatedForm

\u{fc13}: ''

ArabicLigatureThehWithYehIsolatedForm

\u{fc14}: ''

ArabicLigatureJeemWithHahIsolatedForm

\u{fc15}: ''

ArabicLigatureJeemWithMeemIsolatedForm

\u{fc16}: ''

ArabicLigatureHahWithJeemIsolatedForm

\u{fc17}: ''

ArabicLigatureHahWithMeemIsolatedForm

\u{fc18}: ''

ArabicLigatureKhahWithJeemIsolatedForm

\u{fc19}: ''

ArabicLigatureKhahWithHahIsolatedForm

\u{fc1a}: ''

ArabicLigatureKhahWithMeemIsolatedForm

\u{fc1b}: ''

ArabicLigatureSeenWithJeemIsolatedForm

\u{fc1c}: ''

ArabicLigatureSeenWithHahIsolatedForm

\u{fc1d}: ''

ArabicLigatureSeenWithKhahIsolatedForm

\u{fc1e}: ''

ArabicLigatureSeenWithMeemIsolatedForm

\u{fc1f}: ''

ArabicLigatureSadWithHahIsolatedForm

\u{fc20}: ''

ArabicLigatureSadWithMeemIsolatedForm

\u{fc21}: ''

ArabicLigatureDadWithJeemIsolatedForm

\u{fc22}: ''

ArabicLigatureDadWithHahIsolatedForm

\u{fc23}: ''

ArabicLigatureDadWithKhahIsolatedForm

\u{fc24}: ''

ArabicLigatureDadWithMeemIsolatedForm

\u{fc25}: ''

ArabicLigatureTahWithHahIsolatedForm

\u{fc26}: ''

ArabicLigatureTahWithMeemIsolatedForm

\u{fc27}: ''

ArabicLigatureZahWithMeemIsolatedForm

\u{fc28}: ''

ArabicLigatureAinWithJeemIsolatedForm

\u{fc29}: ''

ArabicLigatureAinWithMeemIsolatedForm

\u{fc2a}: ''

ArabicLigatureGhainWithJeemIsolatedForm

\u{fc2b}: ''

ArabicLigatureGhainWithMeemIsolatedForm

\u{fc2c}: ''

ArabicLigatureFehWithJeemIsolatedForm

\u{fc2d}: ''

ArabicLigatureFehWithHahIsolatedForm

\u{fc2e}: ''

ArabicLigatureFehWithKhahIsolatedForm

\u{fc2f}: ''

ArabicLigatureFehWithMeemIsolatedForm

\u{fc30}: ''

ArabicLigatureFehWithAlefMaksuraIsolatedForm

\u{fc31}: ''

ArabicLigatureFehWithYehIsolatedForm

\u{fc32}: ''

ArabicLigatureQafWithHahIsolatedForm

\u{fc33}: ''

ArabicLigatureQafWithMeemIsolatedForm

\u{fc34}: ''

ArabicLigatureQafWithAlefMaksuraIsolatedForm

\u{fc35}: ''

ArabicLigatureQafWithYehIsolatedForm

\u{fc36}: ''

ArabicLigatureKafWithAlefIsolatedForm

\u{fc37}: ''

ArabicLigatureKafWithJeemIsolatedForm

\u{fc38}: ''

ArabicLigatureKafWithHahIsolatedForm

\u{fc39}: ''

ArabicLigatureKafWithKhahIsolatedForm

\u{fc3a}: ''

ArabicLigatureKafWithLamIsolatedForm

\u{fc3b}: ''

ArabicLigatureKafWithMeemIsolatedForm

\u{fc3c}: ''

ArabicLigatureKafWithAlefMaksuraIsolatedForm

\u{fc3d}: ''

ArabicLigatureKafWithYehIsolatedForm

\u{fc3e}: ''

ArabicLigatureLamWithJeemIsolatedForm

\u{fc3f}: ''

ArabicLigatureLamWithHahIsolatedForm

\u{fc40}: ''

ArabicLigatureLamWithKhahIsolatedForm

\u{fc41}: ''

ArabicLigatureLamWithMeemIsolatedForm

\u{fc42}: ''

ArabicLigatureLamWithAlefMaksuraIsolatedForm

\u{fc43}: ''

ArabicLigatureLamWithYehIsolatedForm

\u{fc44}: ''

ArabicLigatureMeemWithJeemIsolatedForm

\u{fc45}: ''

ArabicLigatureMeemWithHahIsolatedForm

\u{fc46}: ''

ArabicLigatureMeemWithKhahIsolatedForm

\u{fc47}: ''

ArabicLigatureMeemWithMeemIsolatedForm

\u{fc48}: ''

ArabicLigatureMeemWithAlefMaksuraIsolatedForm

\u{fc49}: ''

ArabicLigatureMeemWithYehIsolatedForm

\u{fc4a}: ''

ArabicLigatureNoonWithJeemIsolatedForm

\u{fc4b}: ''

ArabicLigatureNoonWithHahIsolatedForm

\u{fc4c}: ''

ArabicLigatureNoonWithKhahIsolatedForm

\u{fc4d}: ''

ArabicLigatureNoonWithMeemIsolatedForm

\u{fc4e}: ''

ArabicLigatureNoonWithAlefMaksuraIsolatedForm

\u{fc4f}: ''

ArabicLigatureNoonWithYehIsolatedForm

\u{fc50}: ''

ArabicLigatureHehWithJeemIsolatedForm

\u{fc51}: ''

ArabicLigatureHehWithMeemIsolatedForm

\u{fc52}: ''

ArabicLigatureHehWithAlefMaksuraIsolatedForm

\u{fc53}: ''

ArabicLigatureHehWithYehIsolatedForm

\u{fc54}: ''

ArabicLigatureYehWithJeemIsolatedForm

\u{fc55}: ''

ArabicLigatureYehWithHahIsolatedForm

\u{fc56}: ''

ArabicLigatureYehWithKhahIsolatedForm

\u{fc57}: ''

ArabicLigatureYehWithMeemIsolatedForm

\u{fc58}: ''

ArabicLigatureYehWithAlefMaksuraIsolatedForm

\u{fc59}: ''

ArabicLigatureYehWithYehIsolatedForm

\u{fc5a}: ''

ArabicLigatureThalWithSuperscriptAlefIsolatedForm

\u{fc5b}: ''

ArabicLigatureRehWithSuperscriptAlefIsolatedForm

\u{fc5c}: ''

ArabicLigatureAlefMaksuraWithSuperscriptAlefIsolatedForm

\u{fc5d}: ''

ArabicLigatureShaddaWithDammatanIsolatedForm

\u{fc5e}: ''

ArabicLigatureShaddaWithKasratanIsolatedForm

\u{fc5f}: ''

ArabicLigatureShaddaWithFathaIsolatedForm

\u{fc60}: ''

ArabicLigatureShaddaWithDammaIsolatedForm

\u{fc61}: ''

ArabicLigatureShaddaWithKasraIsolatedForm

\u{fc62}: ''

ArabicLigatureShaddaWithSuperscriptAlefIsolatedForm

\u{fc63}: ''

ArabicLigatureYehWithHamzaAboveWithRehFinalForm

\u{fc64}: ''

ArabicLigatureYehWithHamzaAboveWithZainFinalForm

\u{fc65}: ''

ArabicLigatureYehWithHamzaAboveWithMeemFinalForm

\u{fc66}: ''

ArabicLigatureYehWithHamzaAboveWithNoonFinalForm

\u{fc67}: ''

ArabicLigatureYehWithHamzaAboveWithAlefMaksuraFinalForm

\u{fc68}: ''

ArabicLigatureYehWithHamzaAboveWithYehFinalForm

\u{fc69}: ''

ArabicLigatureBehWithRehFinalForm

\u{fc6a}: ''

ArabicLigatureBehWithZainFinalForm

\u{fc6b}: ''

ArabicLigatureBehWithMeemFinalForm

\u{fc6c}: ''

ArabicLigatureBehWithNoonFinalForm

\u{fc6d}: ''

ArabicLigatureBehWithAlefMaksuraFinalForm

\u{fc6e}: ''

ArabicLigatureBehWithYehFinalForm

\u{fc6f}: ''

ArabicLigatureTehWithRehFinalForm

\u{fc70}: ''

ArabicLigatureTehWithZainFinalForm

\u{fc71}: ''

ArabicLigatureTehWithMeemFinalForm

\u{fc72}: ''

ArabicLigatureTehWithNoonFinalForm

\u{fc73}: ''

ArabicLigatureTehWithAlefMaksuraFinalForm

\u{fc74}: ''

ArabicLigatureTehWithYehFinalForm

\u{fc75}: ''

ArabicLigatureThehWithRehFinalForm

\u{fc76}: ''

ArabicLigatureThehWithZainFinalForm

\u{fc77}: ''

ArabicLigatureThehWithMeemFinalForm

\u{fc78}: ''

ArabicLigatureThehWithNoonFinalForm

\u{fc79}: ''

ArabicLigatureThehWithAlefMaksuraFinalForm

\u{fc7a}: ''

ArabicLigatureThehWithYehFinalForm

\u{fc7b}: ''

ArabicLigatureFehWithAlefMaksuraFinalForm

\u{fc7c}: ''

ArabicLigatureFehWithYehFinalForm

\u{fc7d}: ''

ArabicLigatureQafWithAlefMaksuraFinalForm

\u{fc7e}: ''

ArabicLigatureQafWithYehFinalForm

\u{fc7f}: ''

ArabicLigatureKafWithAlefFinalForm

\u{fc80}: ''

ArabicLigatureKafWithLamFinalForm

\u{fc81}: ''

ArabicLigatureKafWithMeemFinalForm

\u{fc82}: ''

ArabicLigatureKafWithAlefMaksuraFinalForm

\u{fc83}: ''

ArabicLigatureKafWithYehFinalForm

\u{fc84}: ''

ArabicLigatureLamWithMeemFinalForm

\u{fc85}: ''

ArabicLigatureLamWithAlefMaksuraFinalForm

\u{fc86}: ''

ArabicLigatureLamWithYehFinalForm

\u{fc87}: ''

ArabicLigatureMeemWithAlefFinalForm

\u{fc88}: ''

ArabicLigatureMeemWithMeemFinalForm

\u{fc89}: ''

ArabicLigatureNoonWithRehFinalForm

\u{fc8a}: ''

ArabicLigatureNoonWithZainFinalForm

\u{fc8b}: ''

ArabicLigatureNoonWithMeemFinalForm

\u{fc8c}: ''

ArabicLigatureNoonWithNoonFinalForm

\u{fc8d}: ''

ArabicLigatureNoonWithAlefMaksuraFinalForm

\u{fc8e}: ''

ArabicLigatureNoonWithYehFinalForm

\u{fc8f}: ''

ArabicLigatureAlefMaksuraWithSuperscriptAlefFinalForm

\u{fc90}: ''

ArabicLigatureYehWithRehFinalForm

\u{fc91}: ''

ArabicLigatureYehWithZainFinalForm

\u{fc92}: ''

ArabicLigatureYehWithMeemFinalForm

\u{fc93}: ''

ArabicLigatureYehWithNoonFinalForm

\u{fc94}: ''

ArabicLigatureYehWithAlefMaksuraFinalForm

\u{fc95}: ''

ArabicLigatureYehWithYehFinalForm

\u{fc96}: ''

ArabicLigatureYehWithHamzaAboveWithJeemInitialForm

\u{fc97}: ''

ArabicLigatureYehWithHamzaAboveWithHahInitialForm

\u{fc98}: ''

ArabicLigatureYehWithHamzaAboveWithKhahInitialForm

\u{fc99}: ''

ArabicLigatureYehWithHamzaAboveWithMeemInitialForm

\u{fc9a}: ''

ArabicLigatureYehWithHamzaAboveWithHehInitialForm

\u{fc9b}: ''

ArabicLigatureBehWithJeemInitialForm

\u{fc9c}: ''

ArabicLigatureBehWithHahInitialForm

\u{fc9d}: ''

ArabicLigatureBehWithKhahInitialForm

\u{fc9e}: ''

ArabicLigatureBehWithMeemInitialForm

\u{fc9f}: ''

ArabicLigatureBehWithHehInitialForm

\u{fca0}: ''

ArabicLigatureTehWithJeemInitialForm

\u{fca1}: ''

ArabicLigatureTehWithHahInitialForm

\u{fca2}: ''

ArabicLigatureTehWithKhahInitialForm

\u{fca3}: ''

ArabicLigatureTehWithMeemInitialForm

\u{fca4}: ''

ArabicLigatureTehWithHehInitialForm

\u{fca5}: ''

ArabicLigatureThehWithMeemInitialForm

\u{fca6}: ''

ArabicLigatureJeemWithHahInitialForm

\u{fca7}: ''

ArabicLigatureJeemWithMeemInitialForm

\u{fca8}: ''

ArabicLigatureHahWithJeemInitialForm

\u{fca9}: ''

ArabicLigatureHahWithMeemInitialForm

\u{fcaa}: ''

ArabicLigatureKhahWithJeemInitialForm

\u{fcab}: ''

ArabicLigatureKhahWithMeemInitialForm

\u{fcac}: ''

ArabicLigatureSeenWithJeemInitialForm

\u{fcad}: ''

ArabicLigatureSeenWithHahInitialForm

\u{fcae}: ''

ArabicLigatureSeenWithKhahInitialForm

\u{fcaf}: ''

ArabicLigatureSeenWithMeemInitialForm

\u{fcb0}: ''

ArabicLigatureSadWithHahInitialForm

\u{fcb1}: ''

ArabicLigatureSadWithKhahInitialForm

\u{fcb2}: ''

ArabicLigatureSadWithMeemInitialForm

\u{fcb3}: ''

ArabicLigatureDadWithJeemInitialForm

\u{fcb4}: ''

ArabicLigatureDadWithHahInitialForm

\u{fcb5}: ''

ArabicLigatureDadWithKhahInitialForm

\u{fcb6}: ''

ArabicLigatureDadWithMeemInitialForm

\u{fcb7}: ''

ArabicLigatureTahWithHahInitialForm

\u{fcb8}: ''

ArabicLigatureZahWithMeemInitialForm

\u{fcb9}: ''

ArabicLigatureAinWithJeemInitialForm

\u{fcba}: ''

ArabicLigatureAinWithMeemInitialForm

\u{fcbb}: ''

ArabicLigatureGhainWithJeemInitialForm

\u{fcbc}: ''

ArabicLigatureGhainWithMeemInitialForm

\u{fcbd}: ''

ArabicLigatureFehWithJeemInitialForm

\u{fcbe}: ''

ArabicLigatureFehWithHahInitialForm

\u{fcbf}: ''

ArabicLigatureFehWithKhahInitialForm

\u{fcc0}: ''

ArabicLigatureFehWithMeemInitialForm

\u{fcc1}: ''

ArabicLigatureQafWithHahInitialForm

\u{fcc2}: ''

ArabicLigatureQafWithMeemInitialForm

\u{fcc3}: ''

ArabicLigatureKafWithJeemInitialForm

\u{fcc4}: ''

ArabicLigatureKafWithHahInitialForm

\u{fcc5}: ''

ArabicLigatureKafWithKhahInitialForm

\u{fcc6}: ''

ArabicLigatureKafWithLamInitialForm

\u{fcc7}: ''

ArabicLigatureKafWithMeemInitialForm

\u{fcc8}: ''

ArabicLigatureLamWithJeemInitialForm

\u{fcc9}: ''

ArabicLigatureLamWithHahInitialForm

\u{fcca}: ''

ArabicLigatureLamWithKhahInitialForm

\u{fccb}: ''

ArabicLigatureLamWithMeemInitialForm

\u{fccc}: ''

ArabicLigatureLamWithHehInitialForm

\u{fccd}: ''

ArabicLigatureMeemWithJeemInitialForm

\u{fcce}: ''

ArabicLigatureMeemWithHahInitialForm

\u{fccf}: ''

ArabicLigatureMeemWithKhahInitialForm

\u{fcd0}: ''

ArabicLigatureMeemWithMeemInitialForm

\u{fcd1}: ''

ArabicLigatureNoonWithJeemInitialForm

\u{fcd2}: ''

ArabicLigatureNoonWithHahInitialForm

\u{fcd3}: ''

ArabicLigatureNoonWithKhahInitialForm

\u{fcd4}: ''

ArabicLigatureNoonWithMeemInitialForm

\u{fcd5}: ''

ArabicLigatureNoonWithHehInitialForm

\u{fcd6}: ''

ArabicLigatureHehWithJeemInitialForm

\u{fcd7}: ''

ArabicLigatureHehWithMeemInitialForm

\u{fcd8}: ''

ArabicLigatureHehWithSuperscriptAlefInitialForm

\u{fcd9}: ''

ArabicLigatureYehWithJeemInitialForm

\u{fcda}: ''

ArabicLigatureYehWithHahInitialForm

\u{fcdb}: ''

ArabicLigatureYehWithKhahInitialForm

\u{fcdc}: ''

ArabicLigatureYehWithMeemInitialForm

\u{fcdd}: ''

ArabicLigatureYehWithHehInitialForm

\u{fcde}: ''

ArabicLigatureYehWithHamzaAboveWithMeemMedialForm

\u{fcdf}: ''

ArabicLigatureYehWithHamzaAboveWithHehMedialForm

\u{fce0}: ''

ArabicLigatureBehWithMeemMedialForm

\u{fce1}: ''

ArabicLigatureBehWithHehMedialForm

\u{fce2}: ''

ArabicLigatureTehWithMeemMedialForm

\u{fce3}: ''

ArabicLigatureTehWithHehMedialForm

\u{fce4}: ''

ArabicLigatureThehWithMeemMedialForm

\u{fce5}: ''

ArabicLigatureThehWithHehMedialForm

\u{fce6}: ''

ArabicLigatureSeenWithMeemMedialForm

\u{fce7}: ''

ArabicLigatureSeenWithHehMedialForm

\u{fce8}: ''

ArabicLigatureSheenWithMeemMedialForm

\u{fce9}: ''

ArabicLigatureSheenWithHehMedialForm

\u{fcea}: ''

ArabicLigatureKafWithLamMedialForm

\u{fceb}: ''

ArabicLigatureKafWithMeemMedialForm

\u{fcec}: ''

ArabicLigatureLamWithMeemMedialForm

\u{fced}: ''

ArabicLigatureNoonWithMeemMedialForm

\u{fcee}: ''

ArabicLigatureNoonWithHehMedialForm

\u{fcef}: ''

ArabicLigatureYehWithMeemMedialForm

\u{fcf0}: ''

ArabicLigatureYehWithHehMedialForm

\u{fcf1}: ''

ArabicLigatureShaddaWithFathaMedialForm

\u{fcf2}: ''

ArabicLigatureShaddaWithDammaMedialForm

\u{fcf3}: ''

ArabicLigatureShaddaWithKasraMedialForm

\u{fcf4}: ''

ArabicLigatureTahWithAlefMaksuraIsolatedForm

\u{fcf5}: ''

ArabicLigatureTahWithYehIsolatedForm

\u{fcf6}: ''

ArabicLigatureAinWithAlefMaksuraIsolatedForm

\u{fcf7}: ''

ArabicLigatureAinWithYehIsolatedForm

\u{fcf8}: ''

ArabicLigatureGhainWithAlefMaksuraIsolatedForm

\u{fcf9}: ''

ArabicLigatureGhainWithYehIsolatedForm

\u{fcfa}: ''

ArabicLigatureSeenWithAlefMaksuraIsolatedForm

\u{fcfb}: ''

ArabicLigatureSeenWithYehIsolatedForm

\u{fcfc}: ''

ArabicLigatureSheenWithAlefMaksuraIsolatedForm

\u{fcfd}: ''

ArabicLigatureSheenWithYehIsolatedForm

\u{fcfe}: ''

ArabicLigatureHahWithAlefMaksuraIsolatedForm

\u{fcff}: ''

ArabicLigatureHahWithYehIsolatedForm

\u{fd00}: ''

ArabicLigatureJeemWithAlefMaksuraIsolatedForm

\u{fd01}: ''

ArabicLigatureJeemWithYehIsolatedForm

\u{fd02}: ''

ArabicLigatureKhahWithAlefMaksuraIsolatedForm

\u{fd03}: ''

ArabicLigatureKhahWithYehIsolatedForm

\u{fd04}: ''

ArabicLigatureSadWithAlefMaksuraIsolatedForm

\u{fd05}: ''

ArabicLigatureSadWithYehIsolatedForm

\u{fd06}: ''

ArabicLigatureDadWithAlefMaksuraIsolatedForm

\u{fd07}: ''

ArabicLigatureDadWithYehIsolatedForm

\u{fd08}: ''

ArabicLigatureSheenWithJeemIsolatedForm

\u{fd09}: ''

ArabicLigatureSheenWithHahIsolatedForm

\u{fd0a}: ''

ArabicLigatureSheenWithKhahIsolatedForm

\u{fd0b}: ''

ArabicLigatureSheenWithMeemIsolatedForm

\u{fd0c}: ''

ArabicLigatureSheenWithRehIsolatedForm

\u{fd0d}: ''

ArabicLigatureSeenWithRehIsolatedForm

\u{fd0e}: ''

ArabicLigatureSadWithRehIsolatedForm

\u{fd0f}: ''

ArabicLigatureDadWithRehIsolatedForm

\u{fd10}: ''

ArabicLigatureTahWithAlefMaksuraFinalForm

\u{fd11}: ''

ArabicLigatureTahWithYehFinalForm

\u{fd12}: ''

ArabicLigatureAinWithAlefMaksuraFinalForm

\u{fd13}: ''

ArabicLigatureAinWithYehFinalForm

\u{fd14}: ''

ArabicLigatureGhainWithAlefMaksuraFinalForm

\u{fd15}: ''

ArabicLigatureGhainWithYehFinalForm

\u{fd16}: ''

ArabicLigatureSeenWithAlefMaksuraFinalForm

\u{fd17}: ''

ArabicLigatureSeenWithYehFinalForm

\u{fd18}: ''

ArabicLigatureSheenWithAlefMaksuraFinalForm

\u{fd19}: ''

ArabicLigatureSheenWithYehFinalForm

\u{fd1a}: ''

ArabicLigatureHahWithAlefMaksuraFinalForm

\u{fd1b}: ''

ArabicLigatureHahWithYehFinalForm

\u{fd1c}: ''

ArabicLigatureJeemWithAlefMaksuraFinalForm

\u{fd1d}: ''

ArabicLigatureJeemWithYehFinalForm

\u{fd1e}: ''

ArabicLigatureKhahWithAlefMaksuraFinalForm

\u{fd1f}: ''

ArabicLigatureKhahWithYehFinalForm

\u{fd20}: ''

ArabicLigatureSadWithAlefMaksuraFinalForm

\u{fd21}: ''

ArabicLigatureSadWithYehFinalForm

\u{fd22}: ''

ArabicLigatureDadWithAlefMaksuraFinalForm

\u{fd23}: ''

ArabicLigatureDadWithYehFinalForm

\u{fd24}: ''

ArabicLigatureSheenWithJeemFinalForm

\u{fd25}: ''

ArabicLigatureSheenWithHahFinalForm

\u{fd26}: ''

ArabicLigatureSheenWithKhahFinalForm

\u{fd27}: ''

ArabicLigatureSheenWithMeemFinalForm

\u{fd28}: ''

ArabicLigatureSheenWithRehFinalForm

\u{fd29}: ''

ArabicLigatureSeenWithRehFinalForm

\u{fd2a}: ''

ArabicLigatureSadWithRehFinalForm

\u{fd2b}: ''

ArabicLigatureDadWithRehFinalForm

\u{fd2c}: ''

ArabicLigatureSheenWithJeemInitialForm

\u{fd2d}: ''

ArabicLigatureSheenWithHahInitialForm

\u{fd2e}: ''

ArabicLigatureSheenWithKhahInitialForm

\u{fd2f}: ''

ArabicLigatureSheenWithMeemInitialForm

\u{fd30}: ''

ArabicLigatureSeenWithHehInitialForm

\u{fd31}: ''

ArabicLigatureSheenWithHehInitialForm

\u{fd32}: ''

ArabicLigatureTahWithMeemInitialForm

\u{fd33}: ''

ArabicLigatureSeenWithJeemMedialForm

\u{fd34}: ''

ArabicLigatureSeenWithHahMedialForm

\u{fd35}: ''

ArabicLigatureSeenWithKhahMedialForm

\u{fd36}: ''

ArabicLigatureSheenWithJeemMedialForm

\u{fd37}: ''

ArabicLigatureSheenWithHahMedialForm

\u{fd38}: ''

ArabicLigatureSheenWithKhahMedialForm

\u{fd39}: ''

ArabicLigatureTahWithMeemMedialForm

\u{fd3a}: ''

ArabicLigatureZahWithMeemMedialForm

\u{fd3b}: ''

ArabicLigatureAlefWithFathatanFinalForm

\u{fd3c}: ''

ArabicLigatureAlefWithFathatanIsolatedForm

\u{fd3d}: ''

OrnateLeftParenthesis

\u{fd3e}: ''

OrnateRightParenthesis

\u{fd3f}: ''

ArabicLigatureTehWithJeemWithMeemInitialForm

\u{fd50}: ''

ArabicLigatureTehWithHahWithJeemFinalForm

\u{fd51}: ''

ArabicLigatureTehWithHahWithJeemInitialForm

\u{fd52}: ''

ArabicLigatureTehWithHahWithMeemInitialForm

\u{fd53}: ''

ArabicLigatureTehWithKhahWithMeemInitialForm

\u{fd54}: ''

ArabicLigatureTehWithMeemWithJeemInitialForm

\u{fd55}: ''

ArabicLigatureTehWithMeemWithHahInitialForm

\u{fd56}: ''

ArabicLigatureTehWithMeemWithKhahInitialForm

\u{fd57}: ''

ArabicLigatureJeemWithMeemWithHahFinalForm

\u{fd58}: ''

ArabicLigatureJeemWithMeemWithHahInitialForm

\u{fd59}: ''

ArabicLigatureHahWithMeemWithYehFinalForm

\u{fd5a}: ''

ArabicLigatureHahWithMeemWithAlefMaksuraFinalForm

\u{fd5b}: ''

ArabicLigatureSeenWithHahWithJeemInitialForm

\u{fd5c}: ''

ArabicLigatureSeenWithJeemWithHahInitialForm

\u{fd5d}: ''

ArabicLigatureSeenWithJeemWithAlefMaksuraFinalForm

\u{fd5e}: ''

ArabicLigatureSeenWithMeemWithHahFinalForm

\u{fd5f}: ''

ArabicLigatureSeenWithMeemWithHahInitialForm

\u{fd60}: ''

ArabicLigatureSeenWithMeemWithJeemInitialForm

\u{fd61}: ''

ArabicLigatureSeenWithMeemWithMeemFinalForm

\u{fd62}: ''

ArabicLigatureSeenWithMeemWithMeemInitialForm

\u{fd63}: ''

ArabicLigatureSadWithHahWithHahFinalForm

\u{fd64}: ''

ArabicLigatureSadWithHahWithHahInitialForm

\u{fd65}: ''

ArabicLigatureSadWithMeemWithMeemFinalForm

\u{fd66}: ''

ArabicLigatureSheenWithHahWithMeemFinalForm

\u{fd67}: ''

ArabicLigatureSheenWithHahWithMeemInitialForm

\u{fd68}: ''

ArabicLigatureSheenWithJeemWithYehFinalForm

\u{fd69}: ''

ArabicLigatureSheenWithMeemWithKhahFinalForm

\u{fd6a}: ''

ArabicLigatureSheenWithMeemWithKhahInitialForm

\u{fd6b}: ''

ArabicLigatureSheenWithMeemWithMeemFinalForm

\u{fd6c}: ''

ArabicLigatureSheenWithMeemWithMeemInitialForm

\u{fd6d}: ''

ArabicLigatureDadWithHahWithAlefMaksuraFinalForm

\u{fd6e}: ''

ArabicLigatureDadWithKhahWithMeemFinalForm

\u{fd6f}: ''

ArabicLigatureDadWithKhahWithMeemInitialForm

\u{fd70}: ''

ArabicLigatureTahWithMeemWithHahFinalForm

\u{fd71}: ''

ArabicLigatureTahWithMeemWithHahInitialForm

\u{fd72}: ''

ArabicLigatureTahWithMeemWithMeemInitialForm

\u{fd73}: ''

ArabicLigatureTahWithMeemWithYehFinalForm

\u{fd74}: ''

ArabicLigatureAinWithJeemWithMeemFinalForm

\u{fd75}: ''

ArabicLigatureAinWithMeemWithMeemFinalForm

\u{fd76}: ''

ArabicLigatureAinWithMeemWithMeemInitialForm

\u{fd77}: ''

ArabicLigatureAinWithMeemWithAlefMaksuraFinalForm

\u{fd78}: ''

ArabicLigatureGhainWithMeemWithMeemFinalForm

\u{fd79}: ''

ArabicLigatureGhainWithMeemWithYehFinalForm

\u{fd7a}: ''

ArabicLigatureGhainWithMeemWithAlefMaksuraFinalForm

\u{fd7b}: ''

ArabicLigatureFehWithKhahWithMeemFinalForm

\u{fd7c}: ''

ArabicLigatureFehWithKhahWithMeemInitialForm

\u{fd7d}: ''

ArabicLigatureQafWithMeemWithHahFinalForm

\u{fd7e}: ''

ArabicLigatureQafWithMeemWithMeemFinalForm

\u{fd7f}: ''

ArabicLigatureLamWithHahWithMeemFinalForm

\u{fd80}: ''

ArabicLigatureLamWithHahWithYehFinalForm

\u{fd81}: ''

ArabicLigatureLamWithHahWithAlefMaksuraFinalForm

\u{fd82}: ''

ArabicLigatureLamWithJeemWithJeemInitialForm

\u{fd83}: ''

ArabicLigatureLamWithJeemWithJeemFinalForm

\u{fd84}: ''

ArabicLigatureLamWithKhahWithMeemFinalForm

\u{fd85}: ''

ArabicLigatureLamWithKhahWithMeemInitialForm

\u{fd86}: ''

ArabicLigatureLamWithMeemWithHahFinalForm

\u{fd87}: ''

ArabicLigatureLamWithMeemWithHahInitialForm

\u{fd88}: ''

ArabicLigatureMeemWithHahWithJeemInitialForm

\u{fd89}: ''

ArabicLigatureMeemWithHahWithMeemInitialForm

\u{fd8a}: ''

ArabicLigatureMeemWithHahWithYehFinalForm

\u{fd8b}: ''

ArabicLigatureMeemWithJeemWithHahInitialForm

\u{fd8c}: ''

ArabicLigatureMeemWithJeemWithMeemInitialForm

\u{fd8d}: ''

ArabicLigatureMeemWithKhahWithJeemInitialForm

\u{fd8e}: ''

ArabicLigatureMeemWithKhahWithMeemInitialForm

\u{fd8f}: ''

ArabicLigatureMeemWithJeemWithKhahInitialForm

\u{fd92}: ''

ArabicLigatureHehWithMeemWithJeemInitialForm

\u{fd93}: ''

ArabicLigatureHehWithMeemWithMeemInitialForm

\u{fd94}: ''

ArabicLigatureNoonWithHahWithMeemInitialForm

\u{fd95}: ''

ArabicLigatureNoonWithHahWithAlefMaksuraFinalForm

\u{fd96}: ''

ArabicLigatureNoonWithJeemWithMeemFinalForm

\u{fd97}: ''

ArabicLigatureNoonWithJeemWithMeemInitialForm

\u{fd98}: ''

ArabicLigatureNoonWithJeemWithAlefMaksuraFinalForm

\u{fd99}: ''

ArabicLigatureNoonWithMeemWithYehFinalForm

\u{fd9a}: ''

ArabicLigatureNoonWithMeemWithAlefMaksuraFinalForm

\u{fd9b}: ''

ArabicLigatureYehWithMeemWithMeemFinalForm

\u{fd9c}: ''

ArabicLigatureYehWithMeemWithMeemInitialForm

\u{fd9d}: ''

ArabicLigatureBehWithKhahWithYehFinalForm

\u{fd9e}: ''

ArabicLigatureTehWithJeemWithYehFinalForm

\u{fd9f}: ''

ArabicLigatureTehWithJeemWithAlefMaksuraFinalForm

\u{fda0}: ''

ArabicLigatureTehWithKhahWithYehFinalForm

\u{fda1}: ''

ArabicLigatureTehWithKhahWithAlefMaksuraFinalForm

\u{fda2}: ''

ArabicLigatureTehWithMeemWithYehFinalForm

\u{fda3}: ''

ArabicLigatureTehWithMeemWithAlefMaksuraFinalForm

\u{fda4}: ''

ArabicLigatureJeemWithMeemWithYehFinalForm

\u{fda5}: ''

ArabicLigatureJeemWithHahWithAlefMaksuraFinalForm

\u{fda6}: ''

ArabicLigatureJeemWithMeemWithAlefMaksuraFinalForm

\u{fda7}: ''

ArabicLigatureSeenWithKhahWithAlefMaksuraFinalForm

\u{fda8}: ''

ArabicLigatureSadWithHahWithYehFinalForm

\u{fda9}: ''

ArabicLigatureSheenWithHahWithYehFinalForm

\u{fdaa}: ''

ArabicLigatureDadWithHahWithYehFinalForm

\u{fdab}: ''

ArabicLigatureLamWithJeemWithYehFinalForm

\u{fdac}: ''

ArabicLigatureLamWithMeemWithYehFinalForm

\u{fdad}: ''

ArabicLigatureYehWithHahWithYehFinalForm

\u{fdae}: ''

ArabicLigatureYehWithJeemWithYehFinalForm

\u{fdaf}: ''

ArabicLigatureYehWithMeemWithYehFinalForm

\u{fdb0}: ''

ArabicLigatureMeemWithMeemWithYehFinalForm

\u{fdb1}: ''

ArabicLigatureQafWithMeemWithYehFinalForm

\u{fdb2}: ''

ArabicLigatureNoonWithHahWithYehFinalForm

\u{fdb3}: ''

ArabicLigatureQafWithMeemWithHahInitialForm

\u{fdb4}: ''

ArabicLigatureLamWithHahWithMeemInitialForm

\u{fdb5}: ''

ArabicLigatureAinWithMeemWithYehFinalForm

\u{fdb6}: ''

ArabicLigatureKafWithMeemWithYehFinalForm

\u{fdb7}: ''

ArabicLigatureNoonWithJeemWithHahInitialForm

\u{fdb8}: ''

ArabicLigatureMeemWithKhahWithYehFinalForm

\u{fdb9}: ''

ArabicLigatureLamWithJeemWithMeemInitialForm

\u{fdba}: ''

ArabicLigatureKafWithMeemWithMeemFinalForm

\u{fdbb}: ''

ArabicLigatureLamWithJeemWithMeemFinalForm

\u{fdbc}: ''

ArabicLigatureNoonWithJeemWithHahFinalForm

\u{fdbd}: ''

ArabicLigatureJeemWithHahWithYehFinalForm

\u{fdbe}: ''

ArabicLigatureHahWithJeemWithYehFinalForm

\u{fdbf}: ''

ArabicLigatureMeemWithJeemWithYehFinalForm

\u{fdc0}: ''

ArabicLigatureFehWithMeemWithYehFinalForm

\u{fdc1}: ''

ArabicLigatureBehWithHahWithYehFinalForm

\u{fdc2}: ''

ArabicLigatureKafWithMeemWithMeemInitialForm

\u{fdc3}: ''

ArabicLigatureAinWithJeemWithMeemInitialForm

\u{fdc4}: ''

ArabicLigatureSadWithMeemWithMeemInitialForm

\u{fdc5}: ''

ArabicLigatureSeenWithKhahWithYehFinalForm

\u{fdc6}: ''

ArabicLigatureNoonWithJeemWithYehFinalForm

\u{fdc7}: ''

ArabicLigatureSallaUsedAsKoranicStopSignIsolatedForm

\u{fdf0}: ''

ArabicLigatureQalaUsedAsKoranicStopSignIsolatedForm

\u{fdf1}: ''

ArabicLigatureAllahIsolatedForm

\u{fdf2}: ''

ArabicLigatureAkbarIsolatedForm

\u{fdf3}: ''

ArabicLigatureMohammadIsolatedForm

\u{fdf4}: ''

ArabicLigatureSalamIsolatedForm

\u{fdf5}: ''

ArabicLigatureRasoulIsolatedForm

\u{fdf6}: ''

ArabicLigatureAlayheIsolatedForm

\u{fdf7}: ''

ArabicLigatureWasallamIsolatedForm

\u{fdf8}: ''

ArabicLigatureSallaIsolatedForm

\u{fdf9}: ''

ArabicLigatureSallallahouAlayheWasallam

\u{fdfa}: ''

ArabicLigatureJallajalalouhou

\u{fdfb}: ''

RialSign

\u{fdfc}: ''

ArabicLigatureBismillahArDashRahmanArDashRaheem

\u{fdfd}: ''

Methods

impl ArabicPresentationFormsA

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case