Enum unicode_types::arabic_presentation_forms_b::ArabicPresentationFormsB

pub enum ArabicPresentationFormsB {
    ArabicFathatanIsolatedForm,
    ArabicTatweelWithFathatanAbove,
    ArabicDammatanIsolatedForm,
    ArabicTailFragment,
    ArabicKasratanIsolatedForm,
    ArabicFathaIsolatedForm,
    ArabicFathaMedialForm,
    ArabicDammaIsolatedForm,
    ArabicDammaMedialForm,
    ArabicKasraIsolatedForm,
    ArabicKasraMedialForm,
    ArabicShaddaIsolatedForm,
    ArabicShaddaMedialForm,
    ArabicSukunIsolatedForm,
    ArabicSukunMedialForm,
    ArabicLetterHamzaIsolatedForm,
    ArabicLetterAlefWithMaddaAboveIsolatedForm,
    ArabicLetterAlefWithMaddaAboveFinalForm,
    ArabicLetterAlefWithHamzaAboveIsolatedForm,
    ArabicLetterAlefWithHamzaAboveFinalForm,
    ArabicLetterWawWithHamzaAboveIsolatedForm,
    ArabicLetterWawWithHamzaAboveFinalForm,
    ArabicLetterAlefWithHamzaBelowIsolatedForm,
    ArabicLetterAlefWithHamzaBelowFinalForm,
    ArabicLetterYehWithHamzaAboveIsolatedForm,
    ArabicLetterYehWithHamzaAboveFinalForm,
    ArabicLetterYehWithHamzaAboveInitialForm,
    ArabicLetterYehWithHamzaAboveMedialForm,
    ArabicLetterAlefIsolatedForm,
    ArabicLetterAlefFinalForm,
    ArabicLetterBehIsolatedForm,
    ArabicLetterBehFinalForm,
    ArabicLetterBehInitialForm,
    ArabicLetterBehMedialForm,
    ArabicLetterTehMarbutaIsolatedForm,
    ArabicLetterTehMarbutaFinalForm,
    ArabicLetterTehIsolatedForm,
    ArabicLetterTehFinalForm,
    ArabicLetterTehInitialForm,
    ArabicLetterTehMedialForm,
    ArabicLetterThehIsolatedForm,
    ArabicLetterThehFinalForm,
    ArabicLetterThehInitialForm,
    ArabicLetterThehMedialForm,
    ArabicLetterJeemIsolatedForm,
    ArabicLetterJeemFinalForm,
    ArabicLetterJeemInitialForm,
    ArabicLetterJeemMedialForm,
    ArabicLetterHahIsolatedForm,
    ArabicLetterHahFinalForm,
    ArabicLetterHahInitialForm,
    ArabicLetterHahMedialForm,
    ArabicLetterKhahIsolatedForm,
    ArabicLetterKhahFinalForm,
    ArabicLetterKhahInitialForm,
    ArabicLetterKhahMedialForm,
    ArabicLetterDalIsolatedForm,
    ArabicLetterDalFinalForm,
    ArabicLetterThalIsolatedForm,
    ArabicLetterThalFinalForm,
    ArabicLetterRehIsolatedForm,
    ArabicLetterRehFinalForm,
    ArabicLetterZainIsolatedForm,
    ArabicLetterZainFinalForm,
    ArabicLetterSeenIsolatedForm,
    ArabicLetterSeenFinalForm,
    ArabicLetterSeenInitialForm,
    ArabicLetterSeenMedialForm,
    ArabicLetterSheenIsolatedForm,
    ArabicLetterSheenFinalForm,
    ArabicLetterSheenInitialForm,
    ArabicLetterSheenMedialForm,
    ArabicLetterSadIsolatedForm,
    ArabicLetterSadFinalForm,
    ArabicLetterSadInitialForm,
    ArabicLetterSadMedialForm,
    ArabicLetterDadIsolatedForm,
    ArabicLetterDadFinalForm,
    ArabicLetterDadInitialForm,
    ArabicLetterDadMedialForm,
    ArabicLetterTahIsolatedForm,
    ArabicLetterTahFinalForm,
    ArabicLetterTahInitialForm,
    ArabicLetterTahMedialForm,
    ArabicLetterZahIsolatedForm,
    ArabicLetterZahFinalForm,
    ArabicLetterZahInitialForm,
    ArabicLetterZahMedialForm,
    ArabicLetterAinIsolatedForm,
    ArabicLetterAinFinalForm,
    ArabicLetterAinInitialForm,
    ArabicLetterAinMedialForm,
    ArabicLetterGhainIsolatedForm,
    ArabicLetterGhainFinalForm,
    ArabicLetterGhainInitialForm,
    ArabicLetterGhainMedialForm,
    ArabicLetterFehIsolatedForm,
    ArabicLetterFehFinalForm,
    ArabicLetterFehInitialForm,
    ArabicLetterFehMedialForm,
    ArabicLetterQafIsolatedForm,
    ArabicLetterQafFinalForm,
    ArabicLetterQafInitialForm,
    ArabicLetterQafMedialForm,
    ArabicLetterKafIsolatedForm,
    ArabicLetterKafFinalForm,
    ArabicLetterKafInitialForm,
    ArabicLetterKafMedialForm,
    ArabicLetterLamIsolatedForm,
    ArabicLetterLamFinalForm,
    ArabicLetterLamInitialForm,
    ArabicLetterLamMedialForm,
    ArabicLetterMeemIsolatedForm,
    ArabicLetterMeemFinalForm,
    ArabicLetterMeemInitialForm,
    ArabicLetterMeemMedialForm,
    ArabicLetterNoonIsolatedForm,
    ArabicLetterNoonFinalForm,
    ArabicLetterNoonInitialForm,
    ArabicLetterNoonMedialForm,
    ArabicLetterHehIsolatedForm,
    ArabicLetterHehFinalForm,
    ArabicLetterHehInitialForm,
    ArabicLetterHehMedialForm,
    ArabicLetterWawIsolatedForm,
    ArabicLetterWawFinalForm,
    ArabicLetterAlefMaksuraIsolatedForm,
    ArabicLetterAlefMaksuraFinalForm,
    ArabicLetterYehIsolatedForm,
    ArabicLetterYehFinalForm,
    ArabicLetterYehInitialForm,
    ArabicLetterYehMedialForm,
    ArabicLigatureLamWithAlefWithMaddaAboveIsolatedForm,
    ArabicLigatureLamWithAlefWithMaddaAboveFinalForm,
    ArabicLigatureLamWithAlefWithHamzaAboveIsolatedForm,
    ArabicLigatureLamWithAlefWithHamzaAboveFinalForm,
    ArabicLigatureLamWithAlefWithHamzaBelowIsolatedForm,
    ArabicLigatureLamWithAlefWithHamzaBelowFinalForm,
    ArabicLigatureLamWithAlefIsolatedForm,
    ArabicLigatureLamWithAlefFinalForm,
}

An enum to represent all characters in the ArabicPresentationFormsB block.

Variants
ArabicFathatanIsolatedForm

\u{fe70}: 'ﹰ'

ArabicTatweelWithFathatanAbove

\u{fe71}: 'ﹱ'

ArabicDammatanIsolatedForm

\u{fe72}: 'ﹲ'

ArabicTailFragment

\u{fe73}: 'ﹳ'

ArabicKasratanIsolatedForm

\u{fe74}: 'ﹴ'

ArabicFathaIsolatedForm

\u{fe76}: 'ﹶ'

ArabicFathaMedialForm

\u{fe77}: 'ﹷ'

ArabicDammaIsolatedForm

\u{fe78}: 'ﹸ'

ArabicDammaMedialForm

\u{fe79}: 'ﹹ'

ArabicKasraIsolatedForm

\u{fe7a}: 'ﹺ'

ArabicKasraMedialForm

\u{fe7b}: 'ﹻ'

ArabicShaddaIsolatedForm

\u{fe7c}: 'ﹼ'

ArabicShaddaMedialForm

\u{fe7d}: 'ﹽ'

ArabicSukunIsolatedForm

\u{fe7e}: 'ﹾ'

ArabicSukunMedialForm

\u{fe7f}: 'ﹿ'

ArabicLetterHamzaIsolatedForm

\u{fe80}: 'ﺀ'

ArabicLetterAlefWithMaddaAboveIsolatedForm

\u{fe81}: 'ﺁ'

ArabicLetterAlefWithMaddaAboveFinalForm

\u{fe82}: 'ﺂ'

ArabicLetterAlefWithHamzaAboveIsolatedForm

\u{fe83}: 'ﺃ'

ArabicLetterAlefWithHamzaAboveFinalForm

\u{fe84}: 'ﺄ'

ArabicLetterWawWithHamzaAboveIsolatedForm

\u{fe85}: 'ﺅ'

ArabicLetterWawWithHamzaAboveFinalForm

\u{fe86}: 'ﺆ'

ArabicLetterAlefWithHamzaBelowIsolatedForm

\u{fe87}: 'ﺇ'

ArabicLetterAlefWithHamzaBelowFinalForm

\u{fe88}: 'ﺈ'

ArabicLetterYehWithHamzaAboveIsolatedForm

\u{fe89}: 'ﺉ'

ArabicLetterYehWithHamzaAboveFinalForm

\u{fe8a}: 'ﺊ'

ArabicLetterYehWithHamzaAboveInitialForm

\u{fe8b}: 'ﺋ'

ArabicLetterYehWithHamzaAboveMedialForm

\u{fe8c}: 'ﺌ'

ArabicLetterAlefIsolatedForm

\u{fe8d}: 'ﺍ'

ArabicLetterAlefFinalForm

\u{fe8e}: 'ﺎ'

ArabicLetterBehIsolatedForm

\u{fe8f}: 'ﺏ'

ArabicLetterBehFinalForm

\u{fe90}: 'ﺐ'

ArabicLetterBehInitialForm

\u{fe91}: 'ﺑ'

ArabicLetterBehMedialForm

\u{fe92}: 'ﺒ'

ArabicLetterTehMarbutaIsolatedForm

\u{fe93}: 'ﺓ'

ArabicLetterTehMarbutaFinalForm

\u{fe94}: 'ﺔ'

ArabicLetterTehIsolatedForm

\u{fe95}: 'ﺕ'

ArabicLetterTehFinalForm

\u{fe96}: 'ﺖ'

ArabicLetterTehInitialForm

\u{fe97}: 'ﺗ'

ArabicLetterTehMedialForm

\u{fe98}: 'ﺘ'

ArabicLetterThehIsolatedForm

\u{fe99}: 'ﺙ'

ArabicLetterThehFinalForm

\u{fe9a}: 'ﺚ'

ArabicLetterThehInitialForm

\u{fe9b}: 'ﺛ'

ArabicLetterThehMedialForm

\u{fe9c}: 'ﺜ'

ArabicLetterJeemIsolatedForm

\u{fe9d}: 'ﺝ'

ArabicLetterJeemFinalForm

\u{fe9e}: 'ﺞ'

ArabicLetterJeemInitialForm

\u{fe9f}: 'ﺟ'

ArabicLetterJeemMedialForm

\u{fea0}: 'ﺠ'

ArabicLetterHahIsolatedForm

\u{fea1}: 'ﺡ'

ArabicLetterHahFinalForm

\u{fea2}: 'ﺢ'

ArabicLetterHahInitialForm

\u{fea3}: 'ﺣ'

ArabicLetterHahMedialForm

\u{fea4}: 'ﺤ'

ArabicLetterKhahIsolatedForm

\u{fea5}: 'ﺥ'

ArabicLetterKhahFinalForm

\u{fea6}: 'ﺦ'

ArabicLetterKhahInitialForm

\u{fea7}: 'ﺧ'

ArabicLetterKhahMedialForm

\u{fea8}: 'ﺨ'

ArabicLetterDalIsolatedForm

\u{fea9}: 'ﺩ'

ArabicLetterDalFinalForm

\u{feaa}: 'ﺪ'

ArabicLetterThalIsolatedForm

\u{feab}: 'ﺫ'

ArabicLetterThalFinalForm

\u{feac}: 'ﺬ'

ArabicLetterRehIsolatedForm

\u{fead}: 'ﺭ'

ArabicLetterRehFinalForm

\u{feae}: 'ﺮ'

ArabicLetterZainIsolatedForm

\u{feaf}: 'ﺯ'

ArabicLetterZainFinalForm

\u{feb0}: 'ﺰ'

ArabicLetterSeenIsolatedForm

\u{feb1}: 'ﺱ'

ArabicLetterSeenFinalForm

\u{feb2}: 'ﺲ'

ArabicLetterSeenInitialForm

\u{feb3}: 'ﺳ'

ArabicLetterSeenMedialForm

\u{feb4}: 'ﺴ'

ArabicLetterSheenIsolatedForm

\u{feb5}: 'ﺵ'

ArabicLetterSheenFinalForm

\u{feb6}: 'ﺶ'

ArabicLetterSheenInitialForm

\u{feb7}: 'ﺷ'

ArabicLetterSheenMedialForm

\u{feb8}: 'ﺸ'

ArabicLetterSadIsolatedForm

\u{feb9}: 'ﺹ'

ArabicLetterSadFinalForm

\u{feba}: 'ﺺ'

ArabicLetterSadInitialForm

\u{febb}: 'ﺻ'

ArabicLetterSadMedialForm

\u{febc}: 'ﺼ'

ArabicLetterDadIsolatedForm

\u{febd}: 'ﺽ'

ArabicLetterDadFinalForm

\u{febe}: 'ﺾ'

ArabicLetterDadInitialForm

\u{febf}: 'ﺿ'

ArabicLetterDadMedialForm

\u{fec0}: 'ﻀ'

ArabicLetterTahIsolatedForm

\u{fec1}: 'ﻁ'

ArabicLetterTahFinalForm

\u{fec2}: 'ﻂ'

ArabicLetterTahInitialForm

\u{fec3}: 'ﻃ'

ArabicLetterTahMedialForm

\u{fec4}: 'ﻄ'

ArabicLetterZahIsolatedForm

\u{fec5}: 'ﻅ'

ArabicLetterZahFinalForm

\u{fec6}: 'ﻆ'

ArabicLetterZahInitialForm

\u{fec7}: 'ﻇ'

ArabicLetterZahMedialForm

\u{fec8}: 'ﻈ'

ArabicLetterAinIsolatedForm

\u{fec9}: 'ﻉ'

ArabicLetterAinFinalForm

\u{feca}: 'ﻊ'

ArabicLetterAinInitialForm

\u{fecb}: 'ﻋ'

ArabicLetterAinMedialForm

\u{fecc}: 'ﻌ'

ArabicLetterGhainIsolatedForm

\u{fecd}: 'ﻍ'

ArabicLetterGhainFinalForm

\u{fece}: 'ﻎ'

ArabicLetterGhainInitialForm

\u{fecf}: 'ﻏ'

ArabicLetterGhainMedialForm

\u{fed0}: 'ﻐ'

ArabicLetterFehIsolatedForm

\u{fed1}: 'ﻑ'

ArabicLetterFehFinalForm

\u{fed2}: 'ﻒ'

ArabicLetterFehInitialForm

\u{fed3}: 'ﻓ'

ArabicLetterFehMedialForm

\u{fed4}: 'ﻔ'

ArabicLetterQafIsolatedForm

\u{fed5}: 'ﻕ'

ArabicLetterQafFinalForm

\u{fed6}: 'ﻖ'

ArabicLetterQafInitialForm

\u{fed7}: 'ﻗ'

ArabicLetterQafMedialForm

\u{fed8}: 'ﻘ'

ArabicLetterKafIsolatedForm

\u{fed9}: 'ﻙ'

ArabicLetterKafFinalForm

\u{feda}: 'ﻚ'

ArabicLetterKafInitialForm

\u{fedb}: 'ﻛ'

ArabicLetterKafMedialForm

\u{fedc}: 'ﻜ'

ArabicLetterLamIsolatedForm

\u{fedd}: 'ﻝ'

ArabicLetterLamFinalForm

\u{fede}: 'ﻞ'

ArabicLetterLamInitialForm

\u{fedf}: 'ﻟ'

ArabicLetterLamMedialForm

\u{fee0}: 'ﻠ'

ArabicLetterMeemIsolatedForm

\u{fee1}: 'ﻡ'

ArabicLetterMeemFinalForm

\u{fee2}: 'ﻢ'

ArabicLetterMeemInitialForm

\u{fee3}: 'ﻣ'

ArabicLetterMeemMedialForm

\u{fee4}: 'ﻤ'

ArabicLetterNoonIsolatedForm

\u{fee5}: 'ﻥ'

ArabicLetterNoonFinalForm

\u{fee6}: 'ﻦ'

ArabicLetterNoonInitialForm

\u{fee7}: 'ﻧ'

ArabicLetterNoonMedialForm

\u{fee8}: 'ﻨ'

ArabicLetterHehIsolatedForm

\u{fee9}: 'ﻩ'

ArabicLetterHehFinalForm

\u{feea}: 'ﻪ'

ArabicLetterHehInitialForm

\u{feeb}: 'ﻫ'

ArabicLetterHehMedialForm

\u{feec}: 'ﻬ'

ArabicLetterWawIsolatedForm

\u{feed}: 'ﻭ'

ArabicLetterWawFinalForm

\u{feee}: 'ﻮ'

ArabicLetterAlefMaksuraIsolatedForm

\u{feef}: 'ﻯ'

ArabicLetterAlefMaksuraFinalForm

\u{fef0}: 'ﻰ'

ArabicLetterYehIsolatedForm

\u{fef1}: 'ﻱ'

ArabicLetterYehFinalForm

\u{fef2}: 'ﻲ'

ArabicLetterYehInitialForm

\u{fef3}: 'ﻳ'

ArabicLetterYehMedialForm

\u{fef4}: 'ﻴ'

ArabicLigatureLamWithAlefWithMaddaAboveIsolatedForm

\u{fef5}: 'ﻵ'

ArabicLigatureLamWithAlefWithMaddaAboveFinalForm

\u{fef6}: 'ﻶ'

ArabicLigatureLamWithAlefWithHamzaAboveIsolatedForm

\u{fef7}: 'ﻷ'

ArabicLigatureLamWithAlefWithHamzaAboveFinalForm

\u{fef8}: 'ﻸ'

ArabicLigatureLamWithAlefWithHamzaBelowIsolatedForm

\u{fef9}: 'ﻹ'

ArabicLigatureLamWithAlefWithHamzaBelowFinalForm

\u{fefa}: 'ﻺ'

ArabicLigatureLamWithAlefIsolatedForm

\u{fefb}: 'ﻻ'

ArabicLigatureLamWithAlefFinalForm

\u{fefc}: 'ﻼ'

Methods

impl ArabicPresentationFormsB

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case