Enum unicode_types::armenian::Armenian

pub enum Armenian {
    CapitalLetterAyb,
    CapitalLetterBen,
    CapitalLetterGim,
    CapitalLetterDa,
    CapitalLetterEch,
    CapitalLetterZa,
    CapitalLetterEh,
    CapitalLetterEt,
    CapitalLetterTo,
    CapitalLetterZhe,
    CapitalLetterIni,
    CapitalLetterLiwn,
    CapitalLetterXeh,
    CapitalLetterCa,
    CapitalLetterKen,
    CapitalLetterHo,
    CapitalLetterJa,
    CapitalLetterGhad,
    CapitalLetterCheh,
    CapitalLetterMen,
    CapitalLetterYi,
    CapitalLetterNow,
    CapitalLetterSha,
    CapitalLetterVo,
    CapitalLetterCha,
    CapitalLetterPeh,
    CapitalLetterJheh,
    CapitalLetterRa,
    CapitalLetterSeh,
    CapitalLetterVew,
    CapitalLetterTiwn,
    CapitalLetterReh,
    CapitalLetterCo,
    CapitalLetterYiwn,
    CapitalLetterPiwr,
    CapitalLetterKeh,
    CapitalLetterOh,
    CapitalLetterFeh,
    ModifierLetterLeftHalfRing,
    Apostrophe,
    EmphasisMark,
    ExclamationMark,
    Comma,
    QuestionMark,
    AbbreviationMark,
    SmallLetterTurnedAyb,
    SmallLetterAyb,
    SmallLetterBen,
    SmallLetterGim,
    SmallLetterDa,
    SmallLetterEch,
    SmallLetterZa,
    SmallLetterEh,
    SmallLetterEt,
    SmallLetterTo,
    SmallLetterZhe,
    SmallLetterIni,
    SmallLetterLiwn,
    SmallLetterXeh,
    SmallLetterCa,
    SmallLetterKen,
    SmallLetterHo,
    SmallLetterJa,
    SmallLetterGhad,
    SmallLetterCheh,
    SmallLetterMen,
    SmallLetterYi,
    SmallLetterNow,
    SmallLetterSha,
    SmallLetterVo,
    SmallLetterCha,
    SmallLetterPeh,
    SmallLetterJheh,
    SmallLetterRa,
    SmallLetterSeh,
    SmallLetterVew,
    SmallLetterTiwn,
    SmallLetterReh,
    SmallLetterCo,
    SmallLetterYiwn,
    SmallLetterPiwr,
    SmallLetterKeh,
    SmallLetterOh,
    SmallLetterFeh,
    SmallLigatureEchYiwn,
    SmallLetterYiWithStroke,
    FullStop,
    Hyphen,
    RightDashFacingEternitySign,
    LeftDashFacingEternitySign,
}

An enum to represent all characters in the Armenian block.

Variants
CapitalLetterAyb

\u{531}: 'Ա'

CapitalLetterBen

\u{532}: 'Բ'

CapitalLetterGim

\u{533}: 'Գ'

CapitalLetterDa

\u{534}: 'Դ'

CapitalLetterEch

\u{535}: 'Ե'

CapitalLetterZa

\u{536}: 'Զ'

CapitalLetterEh

\u{537}: 'Է'

CapitalLetterEt

\u{538}: 'Ը'

CapitalLetterTo

\u{539}: 'Թ'

CapitalLetterZhe

\u{53a}: 'Ժ'

CapitalLetterIni

\u{53b}: 'Ի'

CapitalLetterLiwn

\u{53c}: 'Լ'

CapitalLetterXeh

\u{53d}: 'Խ'

CapitalLetterCa

\u{53e}: 'Ծ'

CapitalLetterKen

\u{53f}: 'Կ'

CapitalLetterHo

\u{540}: 'Հ'

CapitalLetterJa

\u{541}: 'Ձ'

CapitalLetterGhad

\u{542}: 'Ղ'

CapitalLetterCheh

\u{543}: 'Ճ'

CapitalLetterMen

\u{544}: 'Մ'

CapitalLetterYi

\u{545}: 'Յ'

CapitalLetterNow

\u{546}: 'Ն'

CapitalLetterSha

\u{547}: 'Շ'

CapitalLetterVo

\u{548}: 'Ո'

CapitalLetterCha

\u{549}: 'Չ'

CapitalLetterPeh

\u{54a}: 'Պ'

CapitalLetterJheh

\u{54b}: 'Ջ'

CapitalLetterRa

\u{54c}: 'Ռ'

CapitalLetterSeh

\u{54d}: 'Ս'

CapitalLetterVew

\u{54e}: 'Վ'

CapitalLetterTiwn

\u{54f}: 'Տ'

CapitalLetterReh

\u{550}: 'Ր'

CapitalLetterCo

\u{551}: 'Ց'

CapitalLetterYiwn

\u{552}: 'Ւ'

CapitalLetterPiwr

\u{553}: 'Փ'

CapitalLetterKeh

\u{554}: 'Ք'

CapitalLetterOh

\u{555}: 'Օ'

CapitalLetterFeh

\u{556}: 'Ֆ'

ModifierLetterLeftHalfRing

\u{559}: 'ՙ'

Apostrophe

\u{55a}: '՚'

EmphasisMark

\u{55b}: '՛'

ExclamationMark

\u{55c}: '՜'

Comma

\u{55d}: '՝'

QuestionMark

\u{55e}: '՞'

AbbreviationMark

\u{55f}: '՟'

SmallLetterTurnedAyb

\u{560}: 'ՠ'

SmallLetterAyb

\u{561}: 'ա'

SmallLetterBen

\u{562}: 'բ'

SmallLetterGim

\u{563}: 'գ'

SmallLetterDa

\u{564}: 'դ'

SmallLetterEch

\u{565}: 'ե'

SmallLetterZa

\u{566}: 'զ'

SmallLetterEh

\u{567}: 'է'

SmallLetterEt

\u{568}: 'ը'

SmallLetterTo

\u{569}: 'թ'

SmallLetterZhe

\u{56a}: 'ժ'

SmallLetterIni

\u{56b}: 'ի'

SmallLetterLiwn

\u{56c}: 'լ'

SmallLetterXeh

\u{56d}: 'խ'

SmallLetterCa

\u{56e}: 'ծ'

SmallLetterKen

\u{56f}: 'կ'

SmallLetterHo

\u{570}: 'հ'

SmallLetterJa

\u{571}: 'ձ'

SmallLetterGhad

\u{572}: 'ղ'

SmallLetterCheh

\u{573}: 'ճ'

SmallLetterMen

\u{574}: 'մ'

SmallLetterYi

\u{575}: 'յ'

SmallLetterNow

\u{576}: 'ն'

SmallLetterSha

\u{577}: 'շ'

SmallLetterVo

\u{578}: 'ո'

SmallLetterCha

\u{579}: 'չ'

SmallLetterPeh

\u{57a}: 'պ'

SmallLetterJheh

\u{57b}: 'ջ'

SmallLetterRa

\u{57c}: 'ռ'

SmallLetterSeh

\u{57d}: 'ս'

SmallLetterVew

\u{57e}: 'վ'

SmallLetterTiwn

\u{57f}: 'տ'

SmallLetterReh

\u{580}: 'ր'

SmallLetterCo

\u{581}: 'ց'

SmallLetterYiwn

\u{582}: 'ւ'

SmallLetterPiwr

\u{583}: 'փ'

SmallLetterKeh

\u{584}: 'ք'

SmallLetterOh

\u{585}: 'օ'

SmallLetterFeh

\u{586}: 'ֆ'

SmallLigatureEchYiwn

\u{587}: 'և'

SmallLetterYiWithStroke

\u{588}: 'ֈ'

FullStop

\u{589}: '։'

Hyphen

\u{58a}: '֊'

RightDashFacingEternitySign

\u{58d}: '֍'

LeftDashFacingEternitySign

\u{58e}: '֎'

Methods

impl Armenian

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case