Enum unicode_types::arrows::Arrows

pub enum Arrows {
    LeftwardsArrow,
    UpwardsArrow,
    RightwardsArrow,
    DownwardsArrow,
    LeftRightArrow,
    UpDownArrow,
    NorthWestArrow,
    NorthEastArrow,
    SouthEastArrow,
    SouthWestArrow,
    LeftwardsArrowWithStroke,
    RightwardsArrowWithStroke,
    LeftwardsWaveArrow,
    RightwardsWaveArrow,
    LeftwardsTwoHeadedArrow,
    UpwardsTwoHeadedArrow,
    RightwardsTwoHeadedArrow,
    DownwardsTwoHeadedArrow,
    LeftwardsArrowWithTail,
    RightwardsArrowWithTail,
    LeftwardsArrowFromBar,
    UpwardsArrowFromBar,
    RightwardsArrowFromBar,
    DownwardsArrowFromBar,
    UpDownArrowWithBase,
    LeftwardsArrowWithHook,
    RightwardsArrowWithHook,
    LeftwardsArrowWithLoop,
    RightwardsArrowWithLoop,
    LeftRightWaveArrow,
    LeftRightArrowWithStroke,
    DownwardsZigzagArrow,
    UpwardsArrowWithTipLeftwards,
    UpwardsArrowWithTipRightwards,
    DownwardsArrowWithTipLeftwards,
    DownwardsArrowWithTipRightwards,
    RightwardsArrowWithCornerDownwards,
    DownwardsArrowWithCornerLeftwards,
    AnticlockwiseTopSemicircleArrow,
    ClockwiseTopSemicircleArrow,
    NorthWestArrowToLongBar,
    LeftwardsArrowToBarOverRightwardsArrowToBar,
    AnticlockwiseOpenCircleArrow,
    ClockwiseOpenCircleArrow,
    LeftwardsHarpoonWithBarbUpwards,
    LeftwardsHarpoonWithBarbDownwards,
    UpwardsHarpoonWithBarbRightwards,
    UpwardsHarpoonWithBarbLeftwards,
    RightwardsHarpoonWithBarbUpwards,
    RightwardsHarpoonWithBarbDownwards,
    DownwardsHarpoonWithBarbRightwards,
    DownwardsHarpoonWithBarbLeftwards,
    RightwardsArrowOverLeftwardsArrow,
    UpwardsArrowLeftwardsOfDownwardsArrow,
    LeftwardsArrowOverRightwardsArrow,
    LeftwardsPaired,
    UpwardsPaired,
    RightwardsPaired,
    DownwardsPaired,
    LeftwardsHarpoonOverRightwardsHarpoon,
    RightwardsHarpoonOverLeftwardsHarpoon,
    LeftwardsDoubleArrowWithStroke,
    LeftRightDoubleArrowWithStroke,
    RightwardsDoubleArrowWithStroke,
    LeftwardsDoubleArrow,
    UpwardsDoubleArrow,
    RightwardsDoubleArrow,
    DownwardsDoubleArrow,
    LeftRightDoubleArrow,
    UpDownDoubleArrow,
    NorthWestDoubleArrow,
    NorthEastDoubleArrow,
    SouthEastDoubleArrow,
    SouthWestDoubleArrow,
    LeftwardsTripleArrow,
    RightwardsTripleArrow,
    LeftwardsSquiggleArrow,
    RightwardsSquiggleArrow,
    UpwardsArrowWithDoubleStroke,
    DownwardsArrowWithDoubleStroke,
    LeftwardsDashedArrow,
    UpwardsDashedArrow,
    RightwardsDashedArrow,
    DownwardsDashedArrow,
    LeftwardsArrowToBar,
    RightwardsArrowToBar,
    LeftwardsWhiteArrow,
    UpwardsWhiteArrow,
    RightwardsWhiteArrow,
    DownwardsWhiteArrow,
    UpwardsWhiteArrowFromBar,
    UpwardsWhiteArrowOnPedestal,
    UpwardsWhiteArrowOnPedestalWithHorizontalBar,
    UpwardsWhiteArrowOnPedestalWithVerticalBar,
    UpwardsWhiteDoubleArrow,
    UpwardsWhiteDoubleArrowOnPedestal,
    RightwardsWhiteArrowFromWall,
    NorthWestArrowToCorner,
    SouthEastArrowToCorner,
    UpDownWhiteArrow,
    RightArrowWithSmallCircle,
    DownwardsArrowLeftwardsOfUpwardsArrow,
    ThreeRightwards,
    LeftwardsArrowWithVerticalStroke,
    RightwardsArrowWithVerticalStroke,
    LeftRightArrowWithVerticalStroke,
    LeftwardsArrowWithDoubleVerticalStroke,
    RightwardsArrowWithDoubleVerticalStroke,
    LeftRightArrowWithDoubleVerticalStroke,
    LeftwardsOpenDashHeadedArrow,
    RightwardsOpenDashHeadedArrow,
}

An enum to represent all characters in the Arrows block.

Variants
LeftwardsArrow

\u{2190}: '←'

UpwardsArrow

\u{2191}: '↑'

RightwardsArrow

\u{2192}: '→'

DownwardsArrow

\u{2193}: '↓'

LeftRightArrow

\u{2194}: '↔'

UpDownArrow

\u{2195}: '↕'

NorthWestArrow

\u{2196}: '↖'

NorthEastArrow

\u{2197}: '↗'

SouthEastArrow

\u{2198}: '↘'

SouthWestArrow

\u{2199}: '↙'

LeftwardsArrowWithStroke

\u{219a}: '↚'

RightwardsArrowWithStroke

\u{219b}: '↛'

LeftwardsWaveArrow

\u{219c}: '↜'

RightwardsWaveArrow

\u{219d}: '↝'

LeftwardsTwoHeadedArrow

\u{219e}: '↞'

UpwardsTwoHeadedArrow

\u{219f}: '↟'

RightwardsTwoHeadedArrow

\u{21a0}: '↠'

DownwardsTwoHeadedArrow

\u{21a1}: '↡'

LeftwardsArrowWithTail

\u{21a2}: '↢'

RightwardsArrowWithTail

\u{21a3}: '↣'

LeftwardsArrowFromBar

\u{21a4}: '↤'

UpwardsArrowFromBar

\u{21a5}: '↥'

RightwardsArrowFromBar

\u{21a6}: '↦'

DownwardsArrowFromBar

\u{21a7}: '↧'

UpDownArrowWithBase

\u{21a8}: '↨'

LeftwardsArrowWithHook

\u{21a9}: '↩'

RightwardsArrowWithHook

\u{21aa}: '↪'

LeftwardsArrowWithLoop

\u{21ab}: '↫'

RightwardsArrowWithLoop

\u{21ac}: '↬'

LeftRightWaveArrow

\u{21ad}: '↭'

LeftRightArrowWithStroke

\u{21ae}: '↮'

DownwardsZigzagArrow

\u{21af}: '↯'

UpwardsArrowWithTipLeftwards

\u{21b0}: '↰'

UpwardsArrowWithTipRightwards

\u{21b1}: '↱'

DownwardsArrowWithTipLeftwards

\u{21b2}: '↲'

DownwardsArrowWithTipRightwards

\u{21b3}: '↳'

RightwardsArrowWithCornerDownwards

\u{21b4}: '↴'

DownwardsArrowWithCornerLeftwards

\u{21b5}: '↵'

AnticlockwiseTopSemicircleArrow

\u{21b6}: '↶'

ClockwiseTopSemicircleArrow

\u{21b7}: '↷'

NorthWestArrowToLongBar

\u{21b8}: '↸'

LeftwardsArrowToBarOverRightwardsArrowToBar

\u{21b9}: '↹'

AnticlockwiseOpenCircleArrow

\u{21ba}: '↺'

ClockwiseOpenCircleArrow

\u{21bb}: '↻'

LeftwardsHarpoonWithBarbUpwards

\u{21bc}: '↼'

LeftwardsHarpoonWithBarbDownwards

\u{21bd}: '↽'

UpwardsHarpoonWithBarbRightwards

\u{21be}: '↾'

UpwardsHarpoonWithBarbLeftwards

\u{21bf}: '↿'

RightwardsHarpoonWithBarbUpwards

\u{21c0}: '⇀'

RightwardsHarpoonWithBarbDownwards

\u{21c1}: '⇁'

DownwardsHarpoonWithBarbRightwards

\u{21c2}: '⇂'

DownwardsHarpoonWithBarbLeftwards

\u{21c3}: '⇃'

RightwardsArrowOverLeftwardsArrow

\u{21c4}: '⇄'

UpwardsArrowLeftwardsOfDownwardsArrow

\u{21c5}: '⇅'

LeftwardsArrowOverRightwardsArrow

\u{21c6}: '⇆'

LeftwardsPaired

\u{21c7}: '⇇'

UpwardsPaired

\u{21c8}: '⇈'

RightwardsPaired

\u{21c9}: '⇉'

DownwardsPaired

\u{21ca}: '⇊'

LeftwardsHarpoonOverRightwardsHarpoon

\u{21cb}: '⇋'

RightwardsHarpoonOverLeftwardsHarpoon

\u{21cc}: '⇌'

LeftwardsDoubleArrowWithStroke

\u{21cd}: '⇍'

LeftRightDoubleArrowWithStroke

\u{21ce}: '⇎'

RightwardsDoubleArrowWithStroke

\u{21cf}: '⇏'

LeftwardsDoubleArrow

\u{21d0}: '⇐'

UpwardsDoubleArrow

\u{21d1}: '⇑'

RightwardsDoubleArrow

\u{21d2}: '⇒'

DownwardsDoubleArrow

\u{21d3}: '⇓'

LeftRightDoubleArrow

\u{21d4}: '⇔'

UpDownDoubleArrow

\u{21d5}: '⇕'

NorthWestDoubleArrow

\u{21d6}: '⇖'

NorthEastDoubleArrow

\u{21d7}: '⇗'

SouthEastDoubleArrow

\u{21d8}: '⇘'

SouthWestDoubleArrow

\u{21d9}: '⇙'

LeftwardsTripleArrow

\u{21da}: '⇚'

RightwardsTripleArrow

\u{21db}: '⇛'

LeftwardsSquiggleArrow

\u{21dc}: '⇜'

RightwardsSquiggleArrow

\u{21dd}: '⇝'

UpwardsArrowWithDoubleStroke

\u{21de}: '⇞'

DownwardsArrowWithDoubleStroke

\u{21df}: '⇟'

LeftwardsDashedArrow

\u{21e0}: '⇠'

UpwardsDashedArrow

\u{21e1}: '⇡'

RightwardsDashedArrow

\u{21e2}: '⇢'

DownwardsDashedArrow

\u{21e3}: '⇣'

LeftwardsArrowToBar

\u{21e4}: '⇤'

RightwardsArrowToBar

\u{21e5}: '⇥'

LeftwardsWhiteArrow

\u{21e6}: '⇦'

UpwardsWhiteArrow

\u{21e7}: '⇧'

RightwardsWhiteArrow

\u{21e8}: '⇨'

DownwardsWhiteArrow

\u{21e9}: '⇩'

UpwardsWhiteArrowFromBar

\u{21ea}: '⇪'

UpwardsWhiteArrowOnPedestal

\u{21eb}: '⇫'

UpwardsWhiteArrowOnPedestalWithHorizontalBar

\u{21ec}: '⇬'

UpwardsWhiteArrowOnPedestalWithVerticalBar

\u{21ed}: '⇭'

UpwardsWhiteDoubleArrow

\u{21ee}: '⇮'

UpwardsWhiteDoubleArrowOnPedestal

\u{21ef}: '⇯'

RightwardsWhiteArrowFromWall

\u{21f0}: '⇰'

NorthWestArrowToCorner

\u{21f1}: '⇱'

SouthEastArrowToCorner

\u{21f2}: '⇲'

UpDownWhiteArrow

\u{21f3}: '⇳'

RightArrowWithSmallCircle

\u{21f4}: '⇴'

DownwardsArrowLeftwardsOfUpwardsArrow

\u{21f5}: '⇵'

ThreeRightwards

\u{21f6}: '⇶'

LeftwardsArrowWithVerticalStroke

\u{21f7}: '⇷'

RightwardsArrowWithVerticalStroke

\u{21f8}: '⇸'

LeftRightArrowWithVerticalStroke

\u{21f9}: '⇹'

LeftwardsArrowWithDoubleVerticalStroke

\u{21fa}: '⇺'

RightwardsArrowWithDoubleVerticalStroke

\u{21fb}: '⇻'

LeftRightArrowWithDoubleVerticalStroke

\u{21fc}: '⇼'

LeftwardsOpenDashHeadedArrow

\u{21fd}: '⇽'

RightwardsOpenDashHeadedArrow

\u{21fe}: '⇾'

Methods

impl Arrows

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case