Enum unicode_types::bamum::Bamum

pub enum Bamum {
    LetterA,
    LetterKa,
    LetterU,
    LetterKu,
    LetterEe,
    LetterRee,
    LetterTae,
    LetterO,
    LetterNyi,
    LetterI,
    LetterLa,
    LetterPa,
    LetterRii,
    LetterRiee,
    LetterLeeee,
    LetterMeeee,
    LetterTaa,
    LetterNdaa,
    LetterNjaem,
    LetterM,
    LetterSuu,
    LetterMu,
    LetterShii,
    LetterSi,
    LetterSheux,
    LetterSeux,
    LetterKyee,
    LetterKet,
    LetterNuae,
    LetterNu,
    LetterNjuae,
    LetterYoq,
    LetterShu,
    LetterYuq,
    LetterYa,
    LetterNsha,
    LetterKeux,
    LetterPeux,
    LetterNjee,
    LetterNtee,
    LetterPue,
    LetterWue,
    LetterPee,
    LetterFee,
    LetterRu,
    LetterLu,
    LetterMi,
    LetterNi,
    LetterReux,
    LetterRae,
    LetterKen,
    LetterNgkwaen,
    LetterNgga,
    LetterNga,
    LetterSho,
    LetterPuae,
    LetterFu,
    LetterFom,
    LetterWa,
    LetterNa,
    LetterLi,
    LetterPi,
    LetterLoq,
    LetterKo,
    LetterMben,
    LetterRen,
    LetterMen,
    LetterMa,
    LetterTi,
    LetterKi,
    LetterMo,
    LetterMbaa,
    LetterTet,
    LetterKpa,
    LetterTen,
    LetterNtuu,
    LetterSamba,
    LetterFaamae,
    LetterKovuu,
    LetterKoghom,
    CombiningMarkKoqndon,
    CombiningMarkTukwentis,
    Njaemli,
    FullStop,
    Colon,
    Comma,
    Semicolon,
    QuestionMark,
}

An enum to represent all characters in the Bamum block.

Variants
LetterA

\u{a6a0}: 'ꚠ'

LetterKa

\u{a6a1}: 'ꚡ'

LetterU

\u{a6a2}: 'ꚢ'

LetterKu

\u{a6a3}: 'ꚣ'

LetterEe

\u{a6a4}: 'ꚤ'

LetterRee

\u{a6a5}: 'ꚥ'

LetterTae

\u{a6a6}: 'ꚦ'

LetterO

\u{a6a7}: 'ꚧ'

LetterNyi

\u{a6a8}: 'ꚨ'

LetterI

\u{a6a9}: 'ꚩ'

LetterLa

\u{a6aa}: 'ꚪ'

LetterPa

\u{a6ab}: 'ꚫ'

LetterRii

\u{a6ac}: 'ꚬ'

LetterRiee

\u{a6ad}: 'ꚭ'

LetterLeeee

\u{a6ae}: 'ꚮ'

LetterMeeee

\u{a6af}: 'ꚯ'

LetterTaa

\u{a6b0}: 'ꚰ'

LetterNdaa

\u{a6b1}: 'ꚱ'

LetterNjaem

\u{a6b2}: 'ꚲ'

LetterM

\u{a6b3}: 'ꚳ'

LetterSuu

\u{a6b4}: 'ꚴ'

LetterMu

\u{a6b5}: 'ꚵ'

LetterShii

\u{a6b6}: 'ꚶ'

LetterSi

\u{a6b7}: 'ꚷ'

LetterSheux

\u{a6b8}: 'ꚸ'

LetterSeux

\u{a6b9}: 'ꚹ'

LetterKyee

\u{a6ba}: 'ꚺ'

LetterKet

\u{a6bb}: 'ꚻ'

LetterNuae

\u{a6bc}: 'ꚼ'

LetterNu

\u{a6bd}: 'ꚽ'

LetterNjuae

\u{a6be}: 'ꚾ'

LetterYoq

\u{a6bf}: 'ꚿ'

LetterShu

\u{a6c0}: 'ꛀ'

LetterYuq

\u{a6c1}: 'ꛁ'

LetterYa

\u{a6c2}: 'ꛂ'

LetterNsha

\u{a6c3}: 'ꛃ'

LetterKeux

\u{a6c4}: 'ꛄ'

LetterPeux

\u{a6c5}: 'ꛅ'

LetterNjee

\u{a6c6}: 'ꛆ'

LetterNtee

\u{a6c7}: 'ꛇ'

LetterPue

\u{a6c8}: 'ꛈ'

LetterWue

\u{a6c9}: 'ꛉ'

LetterPee

\u{a6ca}: 'ꛊ'

LetterFee

\u{a6cb}: 'ꛋ'

LetterRu

\u{a6cc}: 'ꛌ'

LetterLu

\u{a6cd}: 'ꛍ'

LetterMi

\u{a6ce}: 'ꛎ'

LetterNi

\u{a6cf}: 'ꛏ'

LetterReux

\u{a6d0}: 'ꛐ'

LetterRae

\u{a6d1}: 'ꛑ'

LetterKen

\u{a6d2}: 'ꛒ'

LetterNgkwaen

\u{a6d3}: 'ꛓ'

LetterNgga

\u{a6d4}: 'ꛔ'

LetterNga

\u{a6d5}: 'ꛕ'

LetterSho

\u{a6d6}: 'ꛖ'

LetterPuae

\u{a6d7}: 'ꛗ'

LetterFu

\u{a6d8}: 'ꛘ'

LetterFom

\u{a6d9}: 'ꛙ'

LetterWa

\u{a6da}: 'ꛚ'

LetterNa

\u{a6db}: 'ꛛ'

LetterLi

\u{a6dc}: 'ꛜ'

LetterPi

\u{a6dd}: 'ꛝ'

LetterLoq

\u{a6de}: 'ꛞ'

LetterKo

\u{a6df}: 'ꛟ'

LetterMben

\u{a6e0}: 'ꛠ'

LetterRen

\u{a6e1}: 'ꛡ'

LetterMen

\u{a6e2}: 'ꛢ'

LetterMa

\u{a6e3}: 'ꛣ'

LetterTi

\u{a6e4}: 'ꛤ'

LetterKi

\u{a6e5}: 'ꛥ'

LetterMo

\u{a6e6}: 'ꛦ'

LetterMbaa

\u{a6e7}: 'ꛧ'

LetterTet

\u{a6e8}: 'ꛨ'

LetterKpa

\u{a6e9}: 'ꛩ'

LetterTen

\u{a6ea}: 'ꛪ'

LetterNtuu

\u{a6eb}: 'ꛫ'

LetterSamba

\u{a6ec}: 'ꛬ'

LetterFaamae

\u{a6ed}: 'ꛭ'

LetterKovuu

\u{a6ee}: 'ꛮ'

LetterKoghom

\u{a6ef}: 'ꛯ'

CombiningMarkKoqndon

\u{a6f0}: '꛰'

CombiningMarkTukwentis

\u{a6f1}: '꛱'

Njaemli

\u{a6f2}: '꛲'

FullStop

\u{a6f3}: '꛳'

Colon

\u{a6f4}: '꛴'

Comma

\u{a6f5}: '꛵'

Semicolon

\u{a6f6}: '꛶'

QuestionMark

\u{a6f7}: '꛷'

Methods

impl Bamum

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case