Enum unicode_types::bamum_supplement::BamumSupplement

pub enum BamumSupplement {
    BamumLetterPhaseDashANgkueMfon,
    BamumLetterPhaseDashAGbieeFon,
    BamumLetterPhaseDashAPonMfonPipaemgbiee,
    BamumLetterPhaseDashAPonMfonPipaemba,
    BamumLetterPhaseDashANaaMfon,
    BamumLetterPhaseDashAShuenshuet,
    BamumLetterPhaseDashATitaMfon,
    BamumLetterPhaseDashANzaMfon,
    BamumLetterPhaseDashAShindaPaNji,
    BamumLetterPhaseDashAPonPaNjiPipaemgbiee,
    BamumLetterPhaseDashAPonPaNjiPipaemba,
    BamumLetterPhaseDashAMaembgbiee,
    BamumLetterPhaseDashATuMaemba,
    BamumLetterPhaseDashANgangu,
    BamumLetterPhaseDashAMaemveux,
    BamumLetterPhaseDashAMansuae,
    BamumLetterPhaseDashAMveuaengam,
    BamumLetterPhaseDashASeunyam,
    BamumLetterPhaseDashANtoqpen,
    BamumLetterPhaseDashAKeukeutnda,
    BamumLetterPhaseDashANkindi,
    BamumLetterPhaseDashASuu,
    BamumLetterPhaseDashANgkuenzeum,
    BamumLetterPhaseDashALapaq,
    BamumLetterPhaseDashALetKut,
    BamumLetterPhaseDashANtapMfaa,
    BamumLetterPhaseDashAMaekeup,
    BamumLetterPhaseDashAPashae,
    BamumLetterPhaseDashAGheuaerae,
    BamumLetterPhaseDashAPamshae,
    BamumLetterPhaseDashAMonNggeuaet,
    BamumLetterPhaseDashANzunMeut,
    BamumLetterPhaseDashAUYuqNae,
    BamumLetterPhaseDashAGheuaegheuae,
    BamumLetterPhaseDashANtapNtaa,
    BamumLetterPhaseDashASisa,
    BamumLetterPhaseDashAMgbasa,
    BamumLetterPhaseDashAMeunjomndeuq,
    BamumLetterPhaseDashAMoompuq,
    BamumLetterPhaseDashAKafa,
    BamumLetterPhaseDashAPaLeeraewa,
    BamumLetterPhaseDashANdaLeeraewa,
    BamumLetterPhaseDashAPet,
    BamumLetterPhaseDashAMaemkpen,
    BamumLetterPhaseDashANika,
    BamumLetterPhaseDashAPup,
    BamumLetterPhaseDashATuaep,
    BamumLetterPhaseDashALuaep,
    BamumLetterPhaseDashASonjam,
    BamumLetterPhaseDashATeuteuwen,
    BamumLetterPhaseDashAMaenyi,
    BamumLetterPhaseDashAKet,
    BamumLetterPhaseDashANdaanggeuaet,
    BamumLetterPhaseDashAKuoq,
    BamumLetterPhaseDashAMoomeut,
    BamumLetterPhaseDashAShum,
    BamumLetterPhaseDashALommae,
    BamumLetterPhaseDashAFiri,
    BamumLetterPhaseDashARom,
    BamumLetterPhaseDashAKpoq,
    BamumLetterPhaseDashASoq,
    BamumLetterPhaseDashAMapPieet,
    BamumLetterPhaseDashAShirae,
    BamumLetterPhaseDashANtap,
    BamumLetterPhaseDashAShoqNshutYum,
    BamumLetterPhaseDashANyitMongkeuaeq,
    BamumLetterPhaseDashAPaarae,
    BamumLetterPhaseDashANkaarae,
    BamumLetterPhaseDashAUnknown,
    BamumLetterPhaseDashANggen,
    BamumLetterPhaseDashAMaesi,
    BamumLetterPhaseDashANjam,
    BamumLetterPhaseDashAMbanyi,
    BamumLetterPhaseDashANyet,
    BamumLetterPhaseDashATeuaen,
    BamumLetterPhaseDashASot,
    BamumLetterPhaseDashAPaam,
    BamumLetterPhaseDashANshiee,
    BamumLetterPhaseDashAMaem,
    BamumLetterPhaseDashANyi,
    BamumLetterPhaseDashAKaq,
    BamumLetterPhaseDashANsha,
    BamumLetterPhaseDashAVee,
    BamumLetterPhaseDashALu,
    BamumLetterPhaseDashANen,
    BamumLetterPhaseDashANaq,
    BamumLetterPhaseDashAMbaq,
    BamumLetterPhaseDashBNshuet,
    BamumLetterPhaseDashBTuMaemgbiee,
    BamumLetterPhaseDashBSiee,
    BamumLetterPhaseDashBSetTu,
    BamumLetterPhaseDashBLomNteum,
    BamumLetterPhaseDashBMbaMaelee,
    BamumLetterPhaseDashBKieem,
    BamumLetterPhaseDashBYeurae,
    BamumLetterPhaseDashBMbaarae,
    BamumLetterPhaseDashBKam,
    BamumLetterPhaseDashBPeeshi,
    BamumLetterPhaseDashBYafuLeeraewa,
    BamumLetterPhaseDashBLamNshutNyam,
    BamumLetterPhaseDashBNtieeSheuoq,
    BamumLetterPhaseDashBNduNjaa,
    BamumLetterPhaseDashBGheugheuaem,
    BamumLetterPhaseDashBPit,
    BamumLetterPhaseDashBTuNsiee,
    BamumLetterPhaseDashBShetNjaq,
    BamumLetterPhaseDashBSheuaeqtu,
    BamumLetterPhaseDashBMfonTeuaeq,
    BamumLetterPhaseDashBMbitMbaaket,
    BamumLetterPhaseDashBNyiNteum,
    BamumLetterPhaseDashBKeupuq,
    BamumLetterPhaseDashBGheughen,
    BamumLetterPhaseDashBKeuyeux,
    BamumLetterPhaseDashBLaanae,
    BamumLetterPhaseDashBParum,
    BamumLetterPhaseDashBVeum,
    BamumLetterPhaseDashBNgkindiMvop,
    BamumLetterPhaseDashBNggeuMbu,
    BamumLetterPhaseDashBWuaet,
    BamumLetterPhaseDashBSakeuae,
    BamumLetterPhaseDashBTaam,
    BamumLetterPhaseDashBMeuq,
    BamumLetterPhaseDashBNgguoq,
    BamumLetterPhaseDashBNgguoqLarge,
    BamumLetterPhaseDashBMfiyaq,
    BamumLetterPhaseDashBSue,
    BamumLetterPhaseDashBMbeuri,
    BamumLetterPhaseDashBMontieen,
    BamumLetterPhaseDashBNyaemae,
    BamumLetterPhaseDashBPungaam,
    BamumLetterPhaseDashBMeutNggeet,
    BamumLetterPhaseDashBFeux,
    BamumLetterPhaseDashBMbuoq,
    BamumLetterPhaseDashBFee,
    BamumLetterPhaseDashBKeuaem,
    BamumLetterPhaseDashBMaNjeuaena,
    BamumLetterPhaseDashBMaNjuqa,
    BamumLetterPhaseDashBLet,
    BamumLetterPhaseDashBNggaam,
    BamumLetterPhaseDashBNsen,
    BamumLetterPhaseDashBMa,
    BamumLetterPhaseDashBKiq,
    BamumLetterPhaseDashBNgom,
    BamumLetterPhaseDashCNgkueMaemba,
    BamumLetterPhaseDashCNza,
    BamumLetterPhaseDashCYum,
    BamumLetterPhaseDashCWangkuoq,
    BamumLetterPhaseDashCNggen,
    BamumLetterPhaseDashCNdeuaeree,
    BamumLetterPhaseDashCNgkaq,
    BamumLetterPhaseDashCGharae,
    BamumLetterPhaseDashCMbeekeet,
    BamumLetterPhaseDashCGbayi,
    BamumLetterPhaseDashCNyirMkparaqMeun,
    BamumLetterPhaseDashCNtuMbit,
    BamumLetterPhaseDashCMbeum,
    BamumLetterPhaseDashCPirieen,
    BamumLetterPhaseDashCNdombu,
    BamumLetterPhaseDashCMbaaCabbageDashTree,
    BamumLetterPhaseDashCKeusheuaep,
    BamumLetterPhaseDashCGhap,
    BamumLetterPhaseDashCKeukaq,
    BamumLetterPhaseDashCYuMuomae,
    BamumLetterPhaseDashCNzeum,
    BamumLetterPhaseDashCMbue,
    BamumLetterPhaseDashCNseuaen,
    BamumLetterPhaseDashCMbit,
    BamumLetterPhaseDashCYeuq,
    BamumLetterPhaseDashCKparaq,
    BamumLetterPhaseDashCKaa,
    BamumLetterPhaseDashCSeux,
    BamumLetterPhaseDashCNdida,
    BamumLetterPhaseDashCTaashae,
    BamumLetterPhaseDashCNjueq,
    BamumLetterPhaseDashCTitaYue,
    BamumLetterPhaseDashCSuaet,
    BamumLetterPhaseDashCNgguaenNyam,
    BamumLetterPhaseDashCVeux,
    BamumLetterPhaseDashCNansanaq,
    BamumLetterPhaseDashCMaKeuaeri,
    BamumLetterPhaseDashCNtaa,
    BamumLetterPhaseDashCNgguon,
    BamumLetterPhaseDashCLap,
    BamumLetterPhaseDashCMbirieen,
    BamumLetterPhaseDashCMgbasaq,
    BamumLetterPhaseDashCNteungba,
    BamumLetterPhaseDashCTeuteux,
    BamumLetterPhaseDashCNggum,
    BamumLetterPhaseDashCFue,
    BamumLetterPhaseDashCNdeut,
    BamumLetterPhaseDashCNsa,
    BamumLetterPhaseDashCNshaq,
    BamumLetterPhaseDashCBung,
    BamumLetterPhaseDashCVeuaepen,
    BamumLetterPhaseDashCMberae,
    BamumLetterPhaseDashCRu,
    BamumLetterPhaseDashCNjaem,
    BamumLetterPhaseDashCLam,
    BamumLetterPhaseDashCTituaep,
    BamumLetterPhaseDashCNsuotNgom,
    BamumLetterPhaseDashCNjeeee,
    BamumLetterPhaseDashCKet,
    BamumLetterPhaseDashCNggu,
    BamumLetterPhaseDashCMaesi,
    BamumLetterPhaseDashCMbuaem,
    BamumLetterPhaseDashCLu,
    BamumLetterPhaseDashCKut,
    BamumLetterPhaseDashCNjam,
    BamumLetterPhaseDashCNgom,
    BamumLetterPhaseDashCWup,
    BamumLetterPhaseDashCNggueet,
    BamumLetterPhaseDashCNsom,
    BamumLetterPhaseDashCNten,
    BamumLetterPhaseDashCKuopNkaarae,
    BamumLetterPhaseDashCNsun,
    BamumLetterPhaseDashCNdam,
    BamumLetterPhaseDashCMaNsiee,
    BamumLetterPhaseDashCYaa,
    BamumLetterPhaseDashCNdap,
    BamumLetterPhaseDashCShueq,
    BamumLetterPhaseDashCSetfon,
    BamumLetterPhaseDashCMbi,
    BamumLetterPhaseDashCMaemba,
    BamumLetterPhaseDashCMbanyi,
    BamumLetterPhaseDashCKeuseux,
    BamumLetterPhaseDashCMbeux,
    BamumLetterPhaseDashCKeum,
    BamumLetterPhaseDashCMbaaPicket,
    BamumLetterPhaseDashCYuwoq,
    BamumLetterPhaseDashCNjeux,
    BamumLetterPhaseDashCMiee,
    BamumLetterPhaseDashCMuae,
    BamumLetterPhaseDashCShiq,
    BamumLetterPhaseDashCKenLaw,
    BamumLetterPhaseDashCKenFatigue,
    BamumLetterPhaseDashCNgaq,
    BamumLetterPhaseDashCNaq,
    BamumLetterPhaseDashCLiq,
    BamumLetterPhaseDashCPin,
    BamumLetterPhaseDashCPen,
    BamumLetterPhaseDashCTet,
    BamumLetterPhaseDashDMbuo,
    BamumLetterPhaseDashDWap,
    BamumLetterPhaseDashDNji,
    BamumLetterPhaseDashDMfon,
    BamumLetterPhaseDashDNjiee,
    BamumLetterPhaseDashDLiee,
    BamumLetterPhaseDashDNjeut,
    BamumLetterPhaseDashDNshee,
    BamumLetterPhaseDashDNggaamae,
    BamumLetterPhaseDashDNyam,
    BamumLetterPhaseDashDWuaen,
    BamumLetterPhaseDashDNgkun,
    BamumLetterPhaseDashDShee,
    BamumLetterPhaseDashDNgkap,
    BamumLetterPhaseDashDKeuaetmeun,
    BamumLetterPhaseDashDTeut,
    BamumLetterPhaseDashDSheuae,
    BamumLetterPhaseDashDNjap,
    BamumLetterPhaseDashDSue,
    BamumLetterPhaseDashDKet,
    BamumLetterPhaseDashDYaemmae,
    BamumLetterPhaseDashDKuom,
    BamumLetterPhaseDashDSap,
    BamumLetterPhaseDashDMfeut,
    BamumLetterPhaseDashDNdeux,
    BamumLetterPhaseDashDMaleeri,
    BamumLetterPhaseDashDMeut,
    BamumLetterPhaseDashDSeuaeq,
    BamumLetterPhaseDashDYen,
    BamumLetterPhaseDashDNjeuaem,
    BamumLetterPhaseDashDKeuotMbuae,
    BamumLetterPhaseDashDNgkeuri,
    BamumLetterPhaseDashDTu,
    BamumLetterPhaseDashDGhaa,
    BamumLetterPhaseDashDNgkyee,
    BamumLetterPhaseDashDFeufeuaet,
    BamumLetterPhaseDashDNdee,
    BamumLetterPhaseDashDMgbofum,
    BamumLetterPhaseDashDLeuaep,
    BamumLetterPhaseDashDNdon,
    BamumLetterPhaseDashDMoni,
    BamumLetterPhaseDashDMgbeun,
    BamumLetterPhaseDashDPuut,
    BamumLetterPhaseDashDMgbiee,
    BamumLetterPhaseDashDMfo,
    BamumLetterPhaseDashDLum,
    BamumLetterPhaseDashDNsieep,
    BamumLetterPhaseDashDMbaa,
    BamumLetterPhaseDashDKwaet,
    BamumLetterPhaseDashDNyet,
    BamumLetterPhaseDashDTeuaen,
    BamumLetterPhaseDashDSot,
    BamumLetterPhaseDashDYuwoq,
    BamumLetterPhaseDashDKeum,
    BamumLetterPhaseDashDRaem,
    BamumLetterPhaseDashDTeeee,
    BamumLetterPhaseDashDNgkeuaeq,
    BamumLetterPhaseDashDMfeuae,
    BamumLetterPhaseDashDNsieet,
    BamumLetterPhaseDashDKeup,
    BamumLetterPhaseDashDPip,
    BamumLetterPhaseDashDPeutae,
    BamumLetterPhaseDashDNyue,
    BamumLetterPhaseDashDLet,
    BamumLetterPhaseDashDNggaam,
    BamumLetterPhaseDashDMfiee,
    BamumLetterPhaseDashDNggwaen,
    BamumLetterPhaseDashDYuom,
    BamumLetterPhaseDashDPap,
    BamumLetterPhaseDashDYuop,
    BamumLetterPhaseDashDNdam,
    BamumLetterPhaseDashDNteum,
    BamumLetterPhaseDashDSuae,
    BamumLetterPhaseDashDKun,
    BamumLetterPhaseDashDNggeux,
    BamumLetterPhaseDashDNgkiee,
    BamumLetterPhaseDashDTuot,
    BamumLetterPhaseDashDMeun,
    BamumLetterPhaseDashDKuq,
    BamumLetterPhaseDashDNsum,
    BamumLetterPhaseDashDTeun,
    BamumLetterPhaseDashDMaenjet,
    BamumLetterPhaseDashDNggap,
    BamumLetterPhaseDashDLeum,
    BamumLetterPhaseDashDNgguom,
    BamumLetterPhaseDashDNshut,
    BamumLetterPhaseDashDNjueq,
    BamumLetterPhaseDashDGheuae,
    BamumLetterPhaseDashDKu,
    BamumLetterPhaseDashDRenOld,
    BamumLetterPhaseDashDTae,
    BamumLetterPhaseDashDToq,
    BamumLetterPhaseDashDNyi,
    BamumLetterPhaseDashDRii,
    BamumLetterPhaseDashDLeeee,
    BamumLetterPhaseDashDMeeee,
    BamumLetterPhaseDashDM,
    BamumLetterPhaseDashDSuu,
    BamumLetterPhaseDashDMu,
    BamumLetterPhaseDashDShii,
    BamumLetterPhaseDashDSheux,
    BamumLetterPhaseDashDKyee,
    BamumLetterPhaseDashDNu,
    BamumLetterPhaseDashDShu,
    BamumLetterPhaseDashDNtee,
    BamumLetterPhaseDashDPee,
    BamumLetterPhaseDashDNi,
    BamumLetterPhaseDashDShoq,
    BamumLetterPhaseDashDPuq,
    BamumLetterPhaseDashDMvop,
    BamumLetterPhaseDashDLoq,
    BamumLetterPhaseDashDRenMuch,
    BamumLetterPhaseDashDTi,
    BamumLetterPhaseDashDNtuu,
    BamumLetterPhaseDashDMbaaSeven,
    BamumLetterPhaseDashDSaq,
    BamumLetterPhaseDashDFaa,
    BamumLetterPhaseDashENdap,
    BamumLetterPhaseDashEToon,
    BamumLetterPhaseDashEMbeum,
    BamumLetterPhaseDashELap,
    BamumLetterPhaseDashEVom,
    BamumLetterPhaseDashELoon,
    BamumLetterPhaseDashEPaa,
    BamumLetterPhaseDashESom,
    BamumLetterPhaseDashERaq,
    BamumLetterPhaseDashENshuop,
    BamumLetterPhaseDashENdun,
    BamumLetterPhaseDashEPuae,
    BamumLetterPhaseDashETam,
    BamumLetterPhaseDashENgka,
    BamumLetterPhaseDashEKpeux,
    BamumLetterPhaseDashEWuo,
    BamumLetterPhaseDashESee,
    BamumLetterPhaseDashENggeuaet,
    BamumLetterPhaseDashEPaam,
    BamumLetterPhaseDashEToo,
    BamumLetterPhaseDashEKuop,
    BamumLetterPhaseDashELom,
    BamumLetterPhaseDashENshiee,
    BamumLetterPhaseDashENgop,
    BamumLetterPhaseDashEMaem,
    BamumLetterPhaseDashENgkeux,
    BamumLetterPhaseDashENgoq,
    BamumLetterPhaseDashENshue,
    BamumLetterPhaseDashERimgba,
    BamumLetterPhaseDashENjeux,
    BamumLetterPhaseDashEPeem,
    BamumLetterPhaseDashESaa,
    BamumLetterPhaseDashENggurae,
    BamumLetterPhaseDashEMgba,
    BamumLetterPhaseDashEGheux,
    BamumLetterPhaseDashENgkeuaem,
    BamumLetterPhaseDashENjaemli,
    BamumLetterPhaseDashEMap,
    BamumLetterPhaseDashELoot,
    BamumLetterPhaseDashENggeeee,
    BamumLetterPhaseDashENdiq,
    BamumLetterPhaseDashETaenNteum,
    BamumLetterPhaseDashESet,
    BamumLetterPhaseDashEPum,
    BamumLetterPhaseDashENdaaSoftness,
    BamumLetterPhaseDashENgguaeshaeNyam,
    BamumLetterPhaseDashEYiee,
    BamumLetterPhaseDashEGheun,
    BamumLetterPhaseDashETuae,
    BamumLetterPhaseDashEYeuae,
    BamumLetterPhaseDashEPo,
    BamumLetterPhaseDashETumae,
    BamumLetterPhaseDashEKeuae,
    BamumLetterPhaseDashESuaen,
    BamumLetterPhaseDashETeuaeq,
    BamumLetterPhaseDashEVeuae,
    BamumLetterPhaseDashEWeux,
    BamumLetterPhaseDashELaam,
    BamumLetterPhaseDashEPu,
    BamumLetterPhaseDashETaaq,
    BamumLetterPhaseDashEGhaamae,
    BamumLetterPhaseDashENgeureut,
    BamumLetterPhaseDashESheuaeq,
    BamumLetterPhaseDashEMgben,
    BamumLetterPhaseDashEMbee,
    BamumLetterPhaseDashENzaq,
    BamumLetterPhaseDashENkom,
    BamumLetterPhaseDashEGbet,
    BamumLetterPhaseDashETum,
    BamumLetterPhaseDashEKuet,
    BamumLetterPhaseDashEYap,
    BamumLetterPhaseDashENyiCleaver,
    BamumLetterPhaseDashEYit,
    BamumLetterPhaseDashEMfeuq,
    BamumLetterPhaseDashENdiaq,
    BamumLetterPhaseDashEPieeq,
    BamumLetterPhaseDashEYueq,
    BamumLetterPhaseDashELeuaem,
    BamumLetterPhaseDashEFue,
    BamumLetterPhaseDashEGbeux,
    BamumLetterPhaseDashENgkup,
    BamumLetterPhaseDashEKet,
    BamumLetterPhaseDashEMae,
    BamumLetterPhaseDashENgkaami,
    BamumLetterPhaseDashEGhet,
    BamumLetterPhaseDashEFa,
    BamumLetterPhaseDashENtum,
    BamumLetterPhaseDashEPeut,
    BamumLetterPhaseDashEYeum,
    BamumLetterPhaseDashENggeuae,
    BamumLetterPhaseDashENyiBetween,
    BamumLetterPhaseDashENzuq,
    BamumLetterPhaseDashEPoon,
    BamumLetterPhaseDashEMiee,
    BamumLetterPhaseDashEFuet,
    BamumLetterPhaseDashENae,
    BamumLetterPhaseDashEMuae,
    BamumLetterPhaseDashEGheuae,
    BamumLetterPhaseDashEFuI,
    BamumLetterPhaseDashEMvi,
    BamumLetterPhaseDashEPuaq,
    BamumLetterPhaseDashENgkum,
    BamumLetterPhaseDashEKut,
    BamumLetterPhaseDashEPiet,
    BamumLetterPhaseDashENtap,
    BamumLetterPhaseDashEYeuaet,
    BamumLetterPhaseDashENggup,
    BamumLetterPhaseDashEPaPeople,
    BamumLetterPhaseDashEFuCall,
    BamumLetterPhaseDashEFom,
    BamumLetterPhaseDashENjee,
    BamumLetterPhaseDashEA,
    BamumLetterPhaseDashEToq,
    BamumLetterPhaseDashEO,
    BamumLetterPhaseDashEI,
    BamumLetterPhaseDashELaq,
    BamumLetterPhaseDashEPaPlural,
    BamumLetterPhaseDashETaa,
    BamumLetterPhaseDashETaq,
    BamumLetterPhaseDashENdaaMyHouse,
    BamumLetterPhaseDashEShiq,
    BamumLetterPhaseDashEYeux,
    BamumLetterPhaseDashENguae,
    BamumLetterPhaseDashEYuaen,
    BamumLetterPhaseDashEYoqSwimming,
    BamumLetterPhaseDashEYoqCover,
    BamumLetterPhaseDashEYuq,
    BamumLetterPhaseDashEYun,
    BamumLetterPhaseDashEKeux,
    BamumLetterPhaseDashEPeux,
    BamumLetterPhaseDashENjeeEpoch,
    BamumLetterPhaseDashEPue,
    BamumLetterPhaseDashEWue,
    BamumLetterPhaseDashEFee,
    BamumLetterPhaseDashEVee,
    BamumLetterPhaseDashELu,
    BamumLetterPhaseDashEMi,
    BamumLetterPhaseDashEReux,
    BamumLetterPhaseDashERae,
    BamumLetterPhaseDashENguaet,
    BamumLetterPhaseDashENga,
    BamumLetterPhaseDashESho,
    BamumLetterPhaseDashEShoq,
    BamumLetterPhaseDashEFuRemedy,
    BamumLetterPhaseDashENa,
    BamumLetterPhaseDashEPi,
    BamumLetterPhaseDashELoq,
    BamumLetterPhaseDashEKo,
    BamumLetterPhaseDashEMen,
    BamumLetterPhaseDashEMa,
    BamumLetterPhaseDashEMaq,
    BamumLetterPhaseDashETeu,
    BamumLetterPhaseDashEKi,
    BamumLetterPhaseDashEMon,
    BamumLetterPhaseDashETen,
    BamumLetterPhaseDashEFaq,
    BamumLetterPhaseDashEGhom,
    BamumLetterPhaseDashFKa,
    BamumLetterPhaseDashFU,
    BamumLetterPhaseDashFKu,
    BamumLetterPhaseDashFEe,
    BamumLetterPhaseDashFRee,
    BamumLetterPhaseDashFTae,
    BamumLetterPhaseDashFNyi,
    BamumLetterPhaseDashFLa,
    BamumLetterPhaseDashFRii,
    BamumLetterPhaseDashFRiee,
    BamumLetterPhaseDashFMeeee,
    BamumLetterPhaseDashFTaa,
    BamumLetterPhaseDashFNdaa,
    BamumLetterPhaseDashFNjaem,
    BamumLetterPhaseDashFM,
    BamumLetterPhaseDashFSuu,
    BamumLetterPhaseDashFShii,
    BamumLetterPhaseDashFSi,
    BamumLetterPhaseDashFSeux,
    BamumLetterPhaseDashFKyee,
    BamumLetterPhaseDashFKet,
    BamumLetterPhaseDashFNuae,
    BamumLetterPhaseDashFNu,
    BamumLetterPhaseDashFNjuae,
    BamumLetterPhaseDashFYoq,
    BamumLetterPhaseDashFShu,
    BamumLetterPhaseDashFYa,
    BamumLetterPhaseDashFNsha,
    BamumLetterPhaseDashFPeux,
    BamumLetterPhaseDashFNtee,
    BamumLetterPhaseDashFWue,
    BamumLetterPhaseDashFPee,
    BamumLetterPhaseDashFRu,
    BamumLetterPhaseDashFNi,
    BamumLetterPhaseDashFReux,
    BamumLetterPhaseDashFKen,
    BamumLetterPhaseDashFNgkwaen,
    BamumLetterPhaseDashFNgga,
    BamumLetterPhaseDashFSho,
    BamumLetterPhaseDashFPuae,
    BamumLetterPhaseDashFFom,
    BamumLetterPhaseDashFWa,
    BamumLetterPhaseDashFLi,
    BamumLetterPhaseDashFLoq,
    BamumLetterPhaseDashFKo,
    BamumLetterPhaseDashFMben,
    BamumLetterPhaseDashFRen,
    BamumLetterPhaseDashFMa,
    BamumLetterPhaseDashFMo,
    BamumLetterPhaseDashFMbaa,
    BamumLetterPhaseDashFTet,
    BamumLetterPhaseDashFKpa,
    BamumLetterPhaseDashFSamba,
    BamumLetterPhaseDashFVueq,
}

An enum to represent all characters in the BamumSupplement block.

Variants
BamumLetterPhaseDashANgkueMfon

\u{16800}: '𖠀'

BamumLetterPhaseDashAGbieeFon

\u{16801}: '𖠁'

BamumLetterPhaseDashAPonMfonPipaemgbiee

\u{16802}: '𖠂'

BamumLetterPhaseDashAPonMfonPipaemba

\u{16803}: '𖠃'

BamumLetterPhaseDashANaaMfon

\u{16804}: '𖠄'

BamumLetterPhaseDashAShuenshuet

\u{16805}: '𖠅'

BamumLetterPhaseDashATitaMfon

\u{16806}: '𖠆'

BamumLetterPhaseDashANzaMfon

\u{16807}: '𖠇'

BamumLetterPhaseDashAShindaPaNji

\u{16808}: '𖠈'

BamumLetterPhaseDashAPonPaNjiPipaemgbiee

\u{16809}: '𖠉'

BamumLetterPhaseDashAPonPaNjiPipaemba

\u{1680a}: '𖠊'

BamumLetterPhaseDashAMaembgbiee

\u{1680b}: '𖠋'

BamumLetterPhaseDashATuMaemba

\u{1680c}: '𖠌'

BamumLetterPhaseDashANgangu

\u{1680d}: '𖠍'

BamumLetterPhaseDashAMaemveux

\u{1680e}: '𖠎'

BamumLetterPhaseDashAMansuae

\u{1680f}: '𖠏'

BamumLetterPhaseDashAMveuaengam

\u{16810}: '𖠐'

BamumLetterPhaseDashASeunyam

\u{16811}: '𖠑'

BamumLetterPhaseDashANtoqpen

\u{16812}: '𖠒'

BamumLetterPhaseDashAKeukeutnda

\u{16813}: '𖠓'

BamumLetterPhaseDashANkindi

\u{16814}: '𖠔'

BamumLetterPhaseDashASuu

\u{16815}: '𖠕'

BamumLetterPhaseDashANgkuenzeum

\u{16816}: '𖠖'

BamumLetterPhaseDashALapaq

\u{16817}: '𖠗'

BamumLetterPhaseDashALetKut

\u{16818}: '𖠘'

BamumLetterPhaseDashANtapMfaa

\u{16819}: '𖠙'

BamumLetterPhaseDashAMaekeup

\u{1681a}: '𖠚'

BamumLetterPhaseDashAPashae

\u{1681b}: '𖠛'

BamumLetterPhaseDashAGheuaerae

\u{1681c}: '𖠜'

BamumLetterPhaseDashAPamshae

\u{1681d}: '𖠝'

BamumLetterPhaseDashAMonNggeuaet

\u{1681e}: '𖠞'

BamumLetterPhaseDashANzunMeut

\u{1681f}: '𖠟'

BamumLetterPhaseDashAUYuqNae

\u{16820}: '𖠠'

BamumLetterPhaseDashAGheuaegheuae

\u{16821}: '𖠡'

BamumLetterPhaseDashANtapNtaa

\u{16822}: '𖠢'

BamumLetterPhaseDashASisa

\u{16823}: '𖠣'

BamumLetterPhaseDashAMgbasa

\u{16824}: '𖠤'

BamumLetterPhaseDashAMeunjomndeuq

\u{16825}: '𖠥'

BamumLetterPhaseDashAMoompuq

\u{16826}: '𖠦'

BamumLetterPhaseDashAKafa

\u{16827}: '𖠧'

BamumLetterPhaseDashAPaLeeraewa

\u{16828}: '𖠨'

BamumLetterPhaseDashANdaLeeraewa

\u{16829}: '𖠩'

BamumLetterPhaseDashAPet

\u{1682a}: '𖠪'

BamumLetterPhaseDashAMaemkpen

\u{1682b}: '𖠫'

BamumLetterPhaseDashANika

\u{1682c}: '𖠬'

BamumLetterPhaseDashAPup

\u{1682d}: '𖠭'

BamumLetterPhaseDashATuaep

\u{1682e}: '𖠮'

BamumLetterPhaseDashALuaep

\u{1682f}: '𖠯'

BamumLetterPhaseDashASonjam

\u{16830}: '𖠰'

BamumLetterPhaseDashATeuteuwen

\u{16831}: '𖠱'

BamumLetterPhaseDashAMaenyi

\u{16832}: '𖠲'

BamumLetterPhaseDashAKet

\u{16833}: '𖠳'

BamumLetterPhaseDashANdaanggeuaet

\u{16834}: '𖠴'

BamumLetterPhaseDashAKuoq

\u{16835}: '𖠵'

BamumLetterPhaseDashAMoomeut

\u{16836}: '𖠶'

BamumLetterPhaseDashAShum

\u{16837}: '𖠷'

BamumLetterPhaseDashALommae

\u{16838}: '𖠸'

BamumLetterPhaseDashAFiri

\u{16839}: '𖠹'

BamumLetterPhaseDashARom

\u{1683a}: '𖠺'

BamumLetterPhaseDashAKpoq

\u{1683b}: '𖠻'

BamumLetterPhaseDashASoq

\u{1683c}: '𖠼'

BamumLetterPhaseDashAMapPieet

\u{1683d}: '𖠽'

BamumLetterPhaseDashAShirae

\u{1683e}: '𖠾'

BamumLetterPhaseDashANtap

\u{1683f}: '𖠿'

BamumLetterPhaseDashAShoqNshutYum

\u{16840}: '𖡀'

BamumLetterPhaseDashANyitMongkeuaeq

\u{16841}: '𖡁'

BamumLetterPhaseDashAPaarae

\u{16842}: '𖡂'

BamumLetterPhaseDashANkaarae

\u{16843}: '𖡃'

BamumLetterPhaseDashAUnknown

\u{16844}: '𖡄'

BamumLetterPhaseDashANggen

\u{16845}: '𖡅'

BamumLetterPhaseDashAMaesi

\u{16846}: '𖡆'

BamumLetterPhaseDashANjam

\u{16847}: '𖡇'

BamumLetterPhaseDashAMbanyi

\u{16848}: '𖡈'

BamumLetterPhaseDashANyet

\u{16849}: '𖡉'

BamumLetterPhaseDashATeuaen

\u{1684a}: '𖡊'

BamumLetterPhaseDashASot

\u{1684b}: '𖡋'

BamumLetterPhaseDashAPaam

\u{1684c}: '𖡌'

BamumLetterPhaseDashANshiee

\u{1684d}: '𖡍'

BamumLetterPhaseDashAMaem

\u{1684e}: '𖡎'

BamumLetterPhaseDashANyi

\u{1684f}: '𖡏'

BamumLetterPhaseDashAKaq

\u{16850}: '𖡐'

BamumLetterPhaseDashANsha

\u{16851}: '𖡑'

BamumLetterPhaseDashAVee

\u{16852}: '𖡒'

BamumLetterPhaseDashALu

\u{16853}: '𖡓'

BamumLetterPhaseDashANen

\u{16854}: '𖡔'

BamumLetterPhaseDashANaq

\u{16855}: '𖡕'

BamumLetterPhaseDashAMbaq

\u{16856}: '𖡖'

BamumLetterPhaseDashBNshuet

\u{16857}: '𖡗'

BamumLetterPhaseDashBTuMaemgbiee

\u{16858}: '𖡘'

BamumLetterPhaseDashBSiee

\u{16859}: '𖡙'

BamumLetterPhaseDashBSetTu

\u{1685a}: '𖡚'

BamumLetterPhaseDashBLomNteum

\u{1685b}: '𖡛'

BamumLetterPhaseDashBMbaMaelee

\u{1685c}: '𖡜'

BamumLetterPhaseDashBKieem

\u{1685d}: '𖡝'

BamumLetterPhaseDashBYeurae

\u{1685e}: '𖡞'

BamumLetterPhaseDashBMbaarae

\u{1685f}: '𖡟'

BamumLetterPhaseDashBKam

\u{16860}: '𖡠'

BamumLetterPhaseDashBPeeshi

\u{16861}: '𖡡'

BamumLetterPhaseDashBYafuLeeraewa

\u{16862}: '𖡢'

BamumLetterPhaseDashBLamNshutNyam

\u{16863}: '𖡣'

BamumLetterPhaseDashBNtieeSheuoq

\u{16864}: '𖡤'

BamumLetterPhaseDashBNduNjaa

\u{16865}: '𖡥'

BamumLetterPhaseDashBGheugheuaem

\u{16866}: '𖡦'

BamumLetterPhaseDashBPit

\u{16867}: '𖡧'

BamumLetterPhaseDashBTuNsiee

\u{16868}: '𖡨'

BamumLetterPhaseDashBShetNjaq

\u{16869}: '𖡩'

BamumLetterPhaseDashBSheuaeqtu

\u{1686a}: '𖡪'

BamumLetterPhaseDashBMfonTeuaeq

\u{1686b}: '𖡫'

BamumLetterPhaseDashBMbitMbaaket

\u{1686c}: '𖡬'

BamumLetterPhaseDashBNyiNteum

\u{1686d}: '𖡭'

BamumLetterPhaseDashBKeupuq

\u{1686e}: '𖡮'

BamumLetterPhaseDashBGheughen

\u{1686f}: '𖡯'

BamumLetterPhaseDashBKeuyeux

\u{16870}: '𖡰'

BamumLetterPhaseDashBLaanae

\u{16871}: '𖡱'

BamumLetterPhaseDashBParum

\u{16872}: '𖡲'

BamumLetterPhaseDashBVeum

\u{16873}: '𖡳'

BamumLetterPhaseDashBNgkindiMvop

\u{16874}: '𖡴'

BamumLetterPhaseDashBNggeuMbu

\u{16875}: '𖡵'

BamumLetterPhaseDashBWuaet

\u{16876}: '𖡶'

BamumLetterPhaseDashBSakeuae

\u{16877}: '𖡷'

BamumLetterPhaseDashBTaam

\u{16878}: '𖡸'

BamumLetterPhaseDashBMeuq

\u{16879}: '𖡹'

BamumLetterPhaseDashBNgguoq

\u{1687a}: '𖡺'

BamumLetterPhaseDashBNgguoqLarge

\u{1687b}: '𖡻'

BamumLetterPhaseDashBMfiyaq

\u{1687c}: '𖡼'

BamumLetterPhaseDashBSue

\u{1687d}: '𖡽'

BamumLetterPhaseDashBMbeuri

\u{1687e}: '𖡾'

BamumLetterPhaseDashBMontieen

\u{1687f}: '𖡿'

BamumLetterPhaseDashBNyaemae

\u{16880}: '𖢀'

BamumLetterPhaseDashBPungaam

\u{16881}: '𖢁'

BamumLetterPhaseDashBMeutNggeet

\u{16882}: '𖢂'

BamumLetterPhaseDashBFeux

\u{16883}: '𖢃'

BamumLetterPhaseDashBMbuoq

\u{16884}: '𖢄'

BamumLetterPhaseDashBFee

\u{16885}: '𖢅'

BamumLetterPhaseDashBKeuaem

\u{16886}: '𖢆'

BamumLetterPhaseDashBMaNjeuaena

\u{16887}: '𖢇'

BamumLetterPhaseDashBMaNjuqa

\u{16888}: '𖢈'

BamumLetterPhaseDashBLet

\u{16889}: '𖢉'

BamumLetterPhaseDashBNggaam

\u{1688a}: '𖢊'

BamumLetterPhaseDashBNsen

\u{1688b}: '𖢋'

BamumLetterPhaseDashBMa

\u{1688c}: '𖢌'

BamumLetterPhaseDashBKiq

\u{1688d}: '𖢍'

BamumLetterPhaseDashBNgom

\u{1688e}: '𖢎'

BamumLetterPhaseDashCNgkueMaemba

\u{1688f}: '𖢏'

BamumLetterPhaseDashCNza

\u{16890}: '𖢐'

BamumLetterPhaseDashCYum

\u{16891}: '𖢑'

BamumLetterPhaseDashCWangkuoq

\u{16892}: '𖢒'

BamumLetterPhaseDashCNggen

\u{16893}: '𖢓'

BamumLetterPhaseDashCNdeuaeree

\u{16894}: '𖢔'

BamumLetterPhaseDashCNgkaq

\u{16895}: '𖢕'

BamumLetterPhaseDashCGharae

\u{16896}: '𖢖'

BamumLetterPhaseDashCMbeekeet

\u{16897}: '𖢗'

BamumLetterPhaseDashCGbayi

\u{16898}: '𖢘'

BamumLetterPhaseDashCNyirMkparaqMeun

\u{16899}: '𖢙'

BamumLetterPhaseDashCNtuMbit

\u{1689a}: '𖢚'

BamumLetterPhaseDashCMbeum

\u{1689b}: '𖢛'

BamumLetterPhaseDashCPirieen

\u{1689c}: '𖢜'

BamumLetterPhaseDashCNdombu

\u{1689d}: '𖢝'

BamumLetterPhaseDashCMbaaCabbageDashTree

\u{1689e}: '𖢞'

BamumLetterPhaseDashCKeusheuaep

\u{1689f}: '𖢟'

BamumLetterPhaseDashCGhap

\u{168a0}: '𖢠'

BamumLetterPhaseDashCKeukaq

\u{168a1}: '𖢡'

BamumLetterPhaseDashCYuMuomae

\u{168a2}: '𖢢'

BamumLetterPhaseDashCNzeum

\u{168a3}: '𖢣'

BamumLetterPhaseDashCMbue

\u{168a4}: '𖢤'

BamumLetterPhaseDashCNseuaen

\u{168a5}: '𖢥'

BamumLetterPhaseDashCMbit

\u{168a6}: '𖢦'

BamumLetterPhaseDashCYeuq

\u{168a7}: '𖢧'

BamumLetterPhaseDashCKparaq

\u{168a8}: '𖢨'

BamumLetterPhaseDashCKaa

\u{168a9}: '𖢩'

BamumLetterPhaseDashCSeux

\u{168aa}: '𖢪'

BamumLetterPhaseDashCNdida

\u{168ab}: '𖢫'

BamumLetterPhaseDashCTaashae

\u{168ac}: '𖢬'

BamumLetterPhaseDashCNjueq

\u{168ad}: '𖢭'

BamumLetterPhaseDashCTitaYue

\u{168ae}: '𖢮'

BamumLetterPhaseDashCSuaet

\u{168af}: '𖢯'

BamumLetterPhaseDashCNgguaenNyam

\u{168b0}: '𖢰'

BamumLetterPhaseDashCVeux

\u{168b1}: '𖢱'

BamumLetterPhaseDashCNansanaq

\u{168b2}: '𖢲'

BamumLetterPhaseDashCMaKeuaeri

\u{168b3}: '𖢳'

BamumLetterPhaseDashCNtaa

\u{168b4}: '𖢴'

BamumLetterPhaseDashCNgguon

\u{168b5}: '𖢵'

BamumLetterPhaseDashCLap

\u{168b6}: '𖢶'

BamumLetterPhaseDashCMbirieen

\u{168b7}: '𖢷'

BamumLetterPhaseDashCMgbasaq

\u{168b8}: '𖢸'

BamumLetterPhaseDashCNteungba

\u{168b9}: '𖢹'

BamumLetterPhaseDashCTeuteux

\u{168ba}: '𖢺'

BamumLetterPhaseDashCNggum

\u{168bb}: '𖢻'

BamumLetterPhaseDashCFue

\u{168bc}: '𖢼'

BamumLetterPhaseDashCNdeut

\u{168bd}: '𖢽'

BamumLetterPhaseDashCNsa

\u{168be}: '𖢾'

BamumLetterPhaseDashCNshaq

\u{168bf}: '𖢿'

BamumLetterPhaseDashCBung

\u{168c0}: '𖣀'

BamumLetterPhaseDashCVeuaepen

\u{168c1}: '𖣁'

BamumLetterPhaseDashCMberae

\u{168c2}: '𖣂'

BamumLetterPhaseDashCRu

\u{168c3}: '𖣃'

BamumLetterPhaseDashCNjaem

\u{168c4}: '𖣄'

BamumLetterPhaseDashCLam

\u{168c5}: '𖣅'

BamumLetterPhaseDashCTituaep

\u{168c6}: '𖣆'

BamumLetterPhaseDashCNsuotNgom

\u{168c7}: '𖣇'

BamumLetterPhaseDashCNjeeee

\u{168c8}: '𖣈'

BamumLetterPhaseDashCKet

\u{168c9}: '𖣉'

BamumLetterPhaseDashCNggu

\u{168ca}: '𖣊'

BamumLetterPhaseDashCMaesi

\u{168cb}: '𖣋'

BamumLetterPhaseDashCMbuaem

\u{168cc}: '𖣌'

BamumLetterPhaseDashCLu

\u{168cd}: '𖣍'

BamumLetterPhaseDashCKut

\u{168ce}: '𖣎'

BamumLetterPhaseDashCNjam

\u{168cf}: '𖣏'

BamumLetterPhaseDashCNgom

\u{168d0}: '𖣐'

BamumLetterPhaseDashCWup

\u{168d1}: '𖣑'

BamumLetterPhaseDashCNggueet

\u{168d2}: '𖣒'

BamumLetterPhaseDashCNsom

\u{168d3}: '𖣓'

BamumLetterPhaseDashCNten

\u{168d4}: '𖣔'

BamumLetterPhaseDashCKuopNkaarae

\u{168d5}: '𖣕'

BamumLetterPhaseDashCNsun

\u{168d6}: '𖣖'

BamumLetterPhaseDashCNdam

\u{168d7}: '𖣗'

BamumLetterPhaseDashCMaNsiee

\u{168d8}: '𖣘'

BamumLetterPhaseDashCYaa

\u{168d9}: '𖣙'

BamumLetterPhaseDashCNdap

\u{168da}: '𖣚'

BamumLetterPhaseDashCShueq

\u{168db}: '𖣛'

BamumLetterPhaseDashCSetfon

\u{168dc}: '𖣜'

BamumLetterPhaseDashCMbi

\u{168dd}: '𖣝'

BamumLetterPhaseDashCMaemba

\u{168de}: '𖣞'

BamumLetterPhaseDashCMbanyi

\u{168df}: '𖣟'

BamumLetterPhaseDashCKeuseux

\u{168e0}: '𖣠'

BamumLetterPhaseDashCMbeux

\u{168e1}: '𖣡'

BamumLetterPhaseDashCKeum

\u{168e2}: '𖣢'

BamumLetterPhaseDashCMbaaPicket

\u{168e3}: '𖣣'

BamumLetterPhaseDashCYuwoq

\u{168e4}: '𖣤'

BamumLetterPhaseDashCNjeux

\u{168e5}: '𖣥'

BamumLetterPhaseDashCMiee

\u{168e6}: '𖣦'

BamumLetterPhaseDashCMuae

\u{168e7}: '𖣧'

BamumLetterPhaseDashCShiq

\u{168e8}: '𖣨'

BamumLetterPhaseDashCKenLaw

\u{168e9}: '𖣩'

BamumLetterPhaseDashCKenFatigue

\u{168ea}: '𖣪'

BamumLetterPhaseDashCNgaq

\u{168eb}: '𖣫'

BamumLetterPhaseDashCNaq

\u{168ec}: '𖣬'

BamumLetterPhaseDashCLiq

\u{168ed}: '𖣭'

BamumLetterPhaseDashCPin

\u{168ee}: '𖣮'

BamumLetterPhaseDashCPen

\u{168ef}: '𖣯'

BamumLetterPhaseDashCTet

\u{168f0}: '𖣰'

BamumLetterPhaseDashDMbuo

\u{168f1}: '𖣱'

BamumLetterPhaseDashDWap

\u{168f2}: '𖣲'

BamumLetterPhaseDashDNji

\u{168f3}: '𖣳'

BamumLetterPhaseDashDMfon

\u{168f4}: '𖣴'

BamumLetterPhaseDashDNjiee

\u{168f5}: '𖣵'

BamumLetterPhaseDashDLiee

\u{168f6}: '𖣶'

BamumLetterPhaseDashDNjeut

\u{168f7}: '𖣷'

BamumLetterPhaseDashDNshee

\u{168f8}: '𖣸'

BamumLetterPhaseDashDNggaamae

\u{168f9}: '𖣹'

BamumLetterPhaseDashDNyam

\u{168fa}: '𖣺'

BamumLetterPhaseDashDWuaen

\u{168fb}: '𖣻'

BamumLetterPhaseDashDNgkun

\u{168fc}: '𖣼'

BamumLetterPhaseDashDShee

\u{168fd}: '𖣽'

BamumLetterPhaseDashDNgkap

\u{168fe}: '𖣾'

BamumLetterPhaseDashDKeuaetmeun

\u{168ff}: '𖣿'

BamumLetterPhaseDashDTeut

\u{16900}: '𖤀'

BamumLetterPhaseDashDSheuae

\u{16901}: '𖤁'

BamumLetterPhaseDashDNjap

\u{16902}: '𖤂'

BamumLetterPhaseDashDSue

\u{16903}: '𖤃'

BamumLetterPhaseDashDKet

\u{16904}: '𖤄'

BamumLetterPhaseDashDYaemmae

\u{16905}: '𖤅'

BamumLetterPhaseDashDKuom

\u{16906}: '𖤆'

BamumLetterPhaseDashDSap

\u{16907}: '𖤇'

BamumLetterPhaseDashDMfeut

\u{16908}: '𖤈'

BamumLetterPhaseDashDNdeux

\u{16909}: '𖤉'

BamumLetterPhaseDashDMaleeri

\u{1690a}: '𖤊'

BamumLetterPhaseDashDMeut

\u{1690b}: '𖤋'

BamumLetterPhaseDashDSeuaeq

\u{1690c}: '𖤌'

BamumLetterPhaseDashDYen

\u{1690d}: '𖤍'

BamumLetterPhaseDashDNjeuaem

\u{1690e}: '𖤎'

BamumLetterPhaseDashDKeuotMbuae

\u{1690f}: '𖤏'

BamumLetterPhaseDashDNgkeuri

\u{16910}: '𖤐'

BamumLetterPhaseDashDTu

\u{16911}: '𖤑'

BamumLetterPhaseDashDGhaa

\u{16912}: '𖤒'

BamumLetterPhaseDashDNgkyee

\u{16913}: '𖤓'

BamumLetterPhaseDashDFeufeuaet

\u{16914}: '𖤔'

BamumLetterPhaseDashDNdee

\u{16915}: '𖤕'

BamumLetterPhaseDashDMgbofum

\u{16916}: '𖤖'

BamumLetterPhaseDashDLeuaep

\u{16917}: '𖤗'

BamumLetterPhaseDashDNdon

\u{16918}: '𖤘'

BamumLetterPhaseDashDMoni

\u{16919}: '𖤙'

BamumLetterPhaseDashDMgbeun

\u{1691a}: '𖤚'

BamumLetterPhaseDashDPuut

\u{1691b}: '𖤛'

BamumLetterPhaseDashDMgbiee

\u{1691c}: '𖤜'

BamumLetterPhaseDashDMfo

\u{1691d}: '𖤝'

BamumLetterPhaseDashDLum

\u{1691e}: '𖤞'

BamumLetterPhaseDashDNsieep

\u{1691f}: '𖤟'

BamumLetterPhaseDashDMbaa

\u{16920}: '𖤠'

BamumLetterPhaseDashDKwaet

\u{16921}: '𖤡'

BamumLetterPhaseDashDNyet

\u{16922}: '𖤢'

BamumLetterPhaseDashDTeuaen

\u{16923}: '𖤣'

BamumLetterPhaseDashDSot

\u{16924}: '𖤤'

BamumLetterPhaseDashDYuwoq

\u{16925}: '𖤥'

BamumLetterPhaseDashDKeum

\u{16926}: '𖤦'

BamumLetterPhaseDashDRaem

\u{16927}: '𖤧'

BamumLetterPhaseDashDTeeee

\u{16928}: '𖤨'

BamumLetterPhaseDashDNgkeuaeq

\u{16929}: '𖤩'

BamumLetterPhaseDashDMfeuae

\u{1692a}: '𖤪'

BamumLetterPhaseDashDNsieet

\u{1692b}: '𖤫'

BamumLetterPhaseDashDKeup

\u{1692c}: '𖤬'

BamumLetterPhaseDashDPip

\u{1692d}: '𖤭'

BamumLetterPhaseDashDPeutae

\u{1692e}: '𖤮'

BamumLetterPhaseDashDNyue

\u{1692f}: '𖤯'

BamumLetterPhaseDashDLet

\u{16930}: '𖤰'

BamumLetterPhaseDashDNggaam

\u{16931}: '𖤱'

BamumLetterPhaseDashDMfiee

\u{16932}: '𖤲'

BamumLetterPhaseDashDNggwaen

\u{16933}: '𖤳'

BamumLetterPhaseDashDYuom

\u{16934}: '𖤴'

BamumLetterPhaseDashDPap

\u{16935}: '𖤵'

BamumLetterPhaseDashDYuop

\u{16936}: '𖤶'

BamumLetterPhaseDashDNdam

\u{16937}: '𖤷'

BamumLetterPhaseDashDNteum

\u{16938}: '𖤸'

BamumLetterPhaseDashDSuae

\u{16939}: '𖤹'

BamumLetterPhaseDashDKun

\u{1693a}: '𖤺'

BamumLetterPhaseDashDNggeux

\u{1693b}: '𖤻'

BamumLetterPhaseDashDNgkiee

\u{1693c}: '𖤼'

BamumLetterPhaseDashDTuot

\u{1693d}: '𖤽'

BamumLetterPhaseDashDMeun

\u{1693e}: '𖤾'

BamumLetterPhaseDashDKuq

\u{1693f}: '𖤿'

BamumLetterPhaseDashDNsum

\u{16940}: '𖥀'

BamumLetterPhaseDashDTeun

\u{16941}: '𖥁'

BamumLetterPhaseDashDMaenjet

\u{16942}: '𖥂'

BamumLetterPhaseDashDNggap

\u{16943}: '𖥃'

BamumLetterPhaseDashDLeum

\u{16944}: '𖥄'

BamumLetterPhaseDashDNgguom

\u{16945}: '𖥅'

BamumLetterPhaseDashDNshut

\u{16946}: '𖥆'

BamumLetterPhaseDashDNjueq

\u{16947}: '𖥇'

BamumLetterPhaseDashDGheuae

\u{16948}: '𖥈'

BamumLetterPhaseDashDKu

\u{16949}: '𖥉'

BamumLetterPhaseDashDRenOld

\u{1694a}: '𖥊'

BamumLetterPhaseDashDTae

\u{1694b}: '𖥋'

BamumLetterPhaseDashDToq

\u{1694c}: '𖥌'

BamumLetterPhaseDashDNyi

\u{1694d}: '𖥍'

BamumLetterPhaseDashDRii

\u{1694e}: '𖥎'

BamumLetterPhaseDashDLeeee

\u{1694f}: '𖥏'

BamumLetterPhaseDashDMeeee

\u{16950}: '𖥐'

BamumLetterPhaseDashDM

\u{16951}: '𖥑'

BamumLetterPhaseDashDSuu

\u{16952}: '𖥒'

BamumLetterPhaseDashDMu

\u{16953}: '𖥓'

BamumLetterPhaseDashDShii

\u{16954}: '𖥔'

BamumLetterPhaseDashDSheux

\u{16955}: '𖥕'

BamumLetterPhaseDashDKyee

\u{16956}: '𖥖'

BamumLetterPhaseDashDNu

\u{16957}: '𖥗'

BamumLetterPhaseDashDShu

\u{16958}: '𖥘'

BamumLetterPhaseDashDNtee

\u{16959}: '𖥙'

BamumLetterPhaseDashDPee

\u{1695a}: '𖥚'

BamumLetterPhaseDashDNi

\u{1695b}: '𖥛'

BamumLetterPhaseDashDShoq

\u{1695c}: '𖥜'

BamumLetterPhaseDashDPuq

\u{1695d}: '𖥝'

BamumLetterPhaseDashDMvop

\u{1695e}: '𖥞'

BamumLetterPhaseDashDLoq

\u{1695f}: '𖥟'

BamumLetterPhaseDashDRenMuch

\u{16960}: '𖥠'

BamumLetterPhaseDashDTi

\u{16961}: '𖥡'

BamumLetterPhaseDashDNtuu

\u{16962}: '𖥢'

BamumLetterPhaseDashDMbaaSeven

\u{16963}: '𖥣'

BamumLetterPhaseDashDSaq

\u{16964}: '𖥤'

BamumLetterPhaseDashDFaa

\u{16965}: '𖥥'

BamumLetterPhaseDashENdap

\u{16966}: '𖥦'

BamumLetterPhaseDashEToon

\u{16967}: '𖥧'

BamumLetterPhaseDashEMbeum

\u{16968}: '𖥨'

BamumLetterPhaseDashELap

\u{16969}: '𖥩'

BamumLetterPhaseDashEVom

\u{1696a}: '𖥪'

BamumLetterPhaseDashELoon

\u{1696b}: '𖥫'

BamumLetterPhaseDashEPaa

\u{1696c}: '𖥬'

BamumLetterPhaseDashESom

\u{1696d}: '𖥭'

BamumLetterPhaseDashERaq

\u{1696e}: '𖥮'

BamumLetterPhaseDashENshuop

\u{1696f}: '𖥯'

BamumLetterPhaseDashENdun

\u{16970}: '𖥰'

BamumLetterPhaseDashEPuae

\u{16971}: '𖥱'

BamumLetterPhaseDashETam

\u{16972}: '𖥲'

BamumLetterPhaseDashENgka

\u{16973}: '𖥳'

BamumLetterPhaseDashEKpeux

\u{16974}: '𖥴'

BamumLetterPhaseDashEWuo

\u{16975}: '𖥵'

BamumLetterPhaseDashESee

\u{16976}: '𖥶'

BamumLetterPhaseDashENggeuaet

\u{16977}: '𖥷'

BamumLetterPhaseDashEPaam

\u{16978}: '𖥸'

BamumLetterPhaseDashEToo

\u{16979}: '𖥹'

BamumLetterPhaseDashEKuop

\u{1697a}: '𖥺'

BamumLetterPhaseDashELom

\u{1697b}: '𖥻'

BamumLetterPhaseDashENshiee

\u{1697c}: '𖥼'

BamumLetterPhaseDashENgop

\u{1697d}: '𖥽'

BamumLetterPhaseDashEMaem

\u{1697e}: '𖥾'

BamumLetterPhaseDashENgkeux

\u{1697f}: '𖥿'

BamumLetterPhaseDashENgoq

\u{16980}: '𖦀'

BamumLetterPhaseDashENshue

\u{16981}: '𖦁'

BamumLetterPhaseDashERimgba

\u{16982}: '𖦂'

BamumLetterPhaseDashENjeux

\u{16983}: '𖦃'

BamumLetterPhaseDashEPeem

\u{16984}: '𖦄'

BamumLetterPhaseDashESaa

\u{16985}: '𖦅'

BamumLetterPhaseDashENggurae

\u{16986}: '𖦆'

BamumLetterPhaseDashEMgba

\u{16987}: '𖦇'

BamumLetterPhaseDashEGheux

\u{16988}: '𖦈'

BamumLetterPhaseDashENgkeuaem

\u{16989}: '𖦉'

BamumLetterPhaseDashENjaemli

\u{1698a}: '𖦊'

BamumLetterPhaseDashEMap

\u{1698b}: '𖦋'

BamumLetterPhaseDashELoot

\u{1698c}: '𖦌'

BamumLetterPhaseDashENggeeee

\u{1698d}: '𖦍'

BamumLetterPhaseDashENdiq

\u{1698e}: '𖦎'

BamumLetterPhaseDashETaenNteum

\u{1698f}: '𖦏'

BamumLetterPhaseDashESet

\u{16990}: '𖦐'

BamumLetterPhaseDashEPum

\u{16991}: '𖦑'

BamumLetterPhaseDashENdaaSoftness

\u{16992}: '𖦒'

BamumLetterPhaseDashENgguaeshaeNyam

\u{16993}: '𖦓'

BamumLetterPhaseDashEYiee

\u{16994}: '𖦔'

BamumLetterPhaseDashEGheun

\u{16995}: '𖦕'

BamumLetterPhaseDashETuae

\u{16996}: '𖦖'

BamumLetterPhaseDashEYeuae

\u{16997}: '𖦗'

BamumLetterPhaseDashEPo

\u{16998}: '𖦘'

BamumLetterPhaseDashETumae

\u{16999}: '𖦙'

BamumLetterPhaseDashEKeuae

\u{1699a}: '𖦚'

BamumLetterPhaseDashESuaen

\u{1699b}: '𖦛'

BamumLetterPhaseDashETeuaeq

\u{1699c}: '𖦜'

BamumLetterPhaseDashEVeuae

\u{1699d}: '𖦝'

BamumLetterPhaseDashEWeux

\u{1699e}: '𖦞'

BamumLetterPhaseDashELaam

\u{1699f}: '𖦟'

BamumLetterPhaseDashEPu

\u{169a0}: '𖦠'

BamumLetterPhaseDashETaaq

\u{169a1}: '𖦡'

BamumLetterPhaseDashEGhaamae

\u{169a2}: '𖦢'

BamumLetterPhaseDashENgeureut

\u{169a3}: '𖦣'

BamumLetterPhaseDashESheuaeq

\u{169a4}: '𖦤'

BamumLetterPhaseDashEMgben

\u{169a5}: '𖦥'

BamumLetterPhaseDashEMbee

\u{169a6}: '𖦦'

BamumLetterPhaseDashENzaq

\u{169a7}: '𖦧'

BamumLetterPhaseDashENkom

\u{169a8}: '𖦨'

BamumLetterPhaseDashEGbet

\u{169a9}: '𖦩'

BamumLetterPhaseDashETum

\u{169aa}: '𖦪'

BamumLetterPhaseDashEKuet

\u{169ab}: '𖦫'

BamumLetterPhaseDashEYap

\u{169ac}: '𖦬'

BamumLetterPhaseDashENyiCleaver

\u{169ad}: '𖦭'

BamumLetterPhaseDashEYit

\u{169ae}: '𖦮'

BamumLetterPhaseDashEMfeuq

\u{169af}: '𖦯'

BamumLetterPhaseDashENdiaq

\u{169b0}: '𖦰'

BamumLetterPhaseDashEPieeq

\u{169b1}: '𖦱'

BamumLetterPhaseDashEYueq

\u{169b2}: '𖦲'

BamumLetterPhaseDashELeuaem

\u{169b3}: '𖦳'

BamumLetterPhaseDashEFue

\u{169b4}: '𖦴'

BamumLetterPhaseDashEGbeux

\u{169b5}: '𖦵'

BamumLetterPhaseDashENgkup

\u{169b6}: '𖦶'

BamumLetterPhaseDashEKet

\u{169b7}: '𖦷'

BamumLetterPhaseDashEMae

\u{169b8}: '𖦸'

BamumLetterPhaseDashENgkaami

\u{169b9}: '𖦹'

BamumLetterPhaseDashEGhet

\u{169ba}: '𖦺'

BamumLetterPhaseDashEFa

\u{169bb}: '𖦻'

BamumLetterPhaseDashENtum

\u{169bc}: '𖦼'

BamumLetterPhaseDashEPeut

\u{169bd}: '𖦽'

BamumLetterPhaseDashEYeum

\u{169be}: '𖦾'

BamumLetterPhaseDashENggeuae

\u{169bf}: '𖦿'

BamumLetterPhaseDashENyiBetween

\u{169c0}: '𖧀'

BamumLetterPhaseDashENzuq

\u{169c1}: '𖧁'

BamumLetterPhaseDashEPoon

\u{169c2}: '𖧂'

BamumLetterPhaseDashEMiee

\u{169c3}: '𖧃'

BamumLetterPhaseDashEFuet

\u{169c4}: '𖧄'

BamumLetterPhaseDashENae

\u{169c5}: '𖧅'

BamumLetterPhaseDashEMuae

\u{169c6}: '𖧆'

BamumLetterPhaseDashEGheuae

\u{169c7}: '𖧇'

BamumLetterPhaseDashEFuI

\u{169c8}: '𖧈'

BamumLetterPhaseDashEMvi

\u{169c9}: '𖧉'

BamumLetterPhaseDashEPuaq

\u{169ca}: '𖧊'

BamumLetterPhaseDashENgkum

\u{169cb}: '𖧋'

BamumLetterPhaseDashEKut

\u{169cc}: '𖧌'

BamumLetterPhaseDashEPiet

\u{169cd}: '𖧍'

BamumLetterPhaseDashENtap

\u{169ce}: '𖧎'

BamumLetterPhaseDashEYeuaet

\u{169cf}: '𖧏'

BamumLetterPhaseDashENggup

\u{169d0}: '𖧐'

BamumLetterPhaseDashEPaPeople

\u{169d1}: '𖧑'

BamumLetterPhaseDashEFuCall

\u{169d2}: '𖧒'

BamumLetterPhaseDashEFom

\u{169d3}: '𖧓'

BamumLetterPhaseDashENjee

\u{169d4}: '𖧔'

BamumLetterPhaseDashEA

\u{169d5}: '𖧕'

BamumLetterPhaseDashEToq

\u{169d6}: '𖧖'

BamumLetterPhaseDashEO

\u{169d7}: '𖧗'

BamumLetterPhaseDashEI

\u{169d8}: '𖧘'

BamumLetterPhaseDashELaq

\u{169d9}: '𖧙'

BamumLetterPhaseDashEPaPlural

\u{169da}: '𖧚'

BamumLetterPhaseDashETaa

\u{169db}: '𖧛'

BamumLetterPhaseDashETaq

\u{169dc}: '𖧜'

BamumLetterPhaseDashENdaaMyHouse

\u{169dd}: '𖧝'

BamumLetterPhaseDashEShiq

\u{169de}: '𖧞'

BamumLetterPhaseDashEYeux

\u{169df}: '𖧟'

BamumLetterPhaseDashENguae

\u{169e0}: '𖧠'

BamumLetterPhaseDashEYuaen

\u{169e1}: '𖧡'

BamumLetterPhaseDashEYoqSwimming

\u{169e2}: '𖧢'

BamumLetterPhaseDashEYoqCover

\u{169e3}: '𖧣'

BamumLetterPhaseDashEYuq

\u{169e4}: '𖧤'

BamumLetterPhaseDashEYun

\u{169e5}: '𖧥'

BamumLetterPhaseDashEKeux

\u{169e6}: '𖧦'

BamumLetterPhaseDashEPeux

\u{169e7}: '𖧧'

BamumLetterPhaseDashENjeeEpoch

\u{169e8}: '𖧨'

BamumLetterPhaseDashEPue

\u{169e9}: '𖧩'

BamumLetterPhaseDashEWue

\u{169ea}: '𖧪'

BamumLetterPhaseDashEFee

\u{169eb}: '𖧫'

BamumLetterPhaseDashEVee

\u{169ec}: '𖧬'

BamumLetterPhaseDashELu

\u{169ed}: '𖧭'

BamumLetterPhaseDashEMi

\u{169ee}: '𖧮'

BamumLetterPhaseDashEReux

\u{169ef}: '𖧯'

BamumLetterPhaseDashERae

\u{169f0}: '𖧰'

BamumLetterPhaseDashENguaet

\u{169f1}: '𖧱'

BamumLetterPhaseDashENga

\u{169f2}: '𖧲'

BamumLetterPhaseDashESho

\u{169f3}: '𖧳'

BamumLetterPhaseDashEShoq

\u{169f4}: '𖧴'

BamumLetterPhaseDashEFuRemedy

\u{169f5}: '𖧵'

BamumLetterPhaseDashENa

\u{169f6}: '𖧶'

BamumLetterPhaseDashEPi

\u{169f7}: '𖧷'

BamumLetterPhaseDashELoq

\u{169f8}: '𖧸'

BamumLetterPhaseDashEKo

\u{169f9}: '𖧹'

BamumLetterPhaseDashEMen

\u{169fa}: '𖧺'

BamumLetterPhaseDashEMa

\u{169fb}: '𖧻'

BamumLetterPhaseDashEMaq

\u{169fc}: '𖧼'

BamumLetterPhaseDashETeu

\u{169fd}: '𖧽'

BamumLetterPhaseDashEKi

\u{169fe}: '𖧾'

BamumLetterPhaseDashEMon

\u{169ff}: '𖧿'

BamumLetterPhaseDashETen

\u{16a00}: '𖨀'

BamumLetterPhaseDashEFaq

\u{16a01}: '𖨁'

BamumLetterPhaseDashEGhom

\u{16a02}: '𖨂'

BamumLetterPhaseDashFKa

\u{16a03}: '𖨃'

BamumLetterPhaseDashFU

\u{16a04}: '𖨄'

BamumLetterPhaseDashFKu

\u{16a05}: '𖨅'

BamumLetterPhaseDashFEe

\u{16a06}: '𖨆'

BamumLetterPhaseDashFRee

\u{16a07}: '𖨇'

BamumLetterPhaseDashFTae

\u{16a08}: '𖨈'

BamumLetterPhaseDashFNyi

\u{16a09}: '𖨉'

BamumLetterPhaseDashFLa

\u{16a0a}: '𖨊'

BamumLetterPhaseDashFRii

\u{16a0b}: '𖨋'

BamumLetterPhaseDashFRiee

\u{16a0c}: '𖨌'

BamumLetterPhaseDashFMeeee

\u{16a0d}: '𖨍'

BamumLetterPhaseDashFTaa

\u{16a0e}: '𖨎'

BamumLetterPhaseDashFNdaa

\u{16a0f}: '𖨏'

BamumLetterPhaseDashFNjaem

\u{16a10}: '𖨐'

BamumLetterPhaseDashFM

\u{16a11}: '𖨑'

BamumLetterPhaseDashFSuu

\u{16a12}: '𖨒'

BamumLetterPhaseDashFShii

\u{16a13}: '𖨓'

BamumLetterPhaseDashFSi

\u{16a14}: '𖨔'

BamumLetterPhaseDashFSeux

\u{16a15}: '𖨕'

BamumLetterPhaseDashFKyee

\u{16a16}: '𖨖'

BamumLetterPhaseDashFKet

\u{16a17}: '𖨗'

BamumLetterPhaseDashFNuae

\u{16a18}: '𖨘'

BamumLetterPhaseDashFNu

\u{16a19}: '𖨙'

BamumLetterPhaseDashFNjuae

\u{16a1a}: '𖨚'

BamumLetterPhaseDashFYoq

\u{16a1b}: '𖨛'

BamumLetterPhaseDashFShu

\u{16a1c}: '𖨜'

BamumLetterPhaseDashFYa

\u{16a1d}: '𖨝'

BamumLetterPhaseDashFNsha

\u{16a1e}: '𖨞'

BamumLetterPhaseDashFPeux

\u{16a1f}: '𖨟'

BamumLetterPhaseDashFNtee

\u{16a20}: '𖨠'

BamumLetterPhaseDashFWue

\u{16a21}: '𖨡'

BamumLetterPhaseDashFPee

\u{16a22}: '𖨢'

BamumLetterPhaseDashFRu

\u{16a23}: '𖨣'

BamumLetterPhaseDashFNi

\u{16a24}: '𖨤'

BamumLetterPhaseDashFReux

\u{16a25}: '𖨥'

BamumLetterPhaseDashFKen

\u{16a26}: '𖨦'

BamumLetterPhaseDashFNgkwaen

\u{16a27}: '𖨧'

BamumLetterPhaseDashFNgga

\u{16a28}: '𖨨'

BamumLetterPhaseDashFSho

\u{16a29}: '𖨩'

BamumLetterPhaseDashFPuae

\u{16a2a}: '𖨪'

BamumLetterPhaseDashFFom

\u{16a2b}: '𖨫'

BamumLetterPhaseDashFWa

\u{16a2c}: '𖨬'

BamumLetterPhaseDashFLi

\u{16a2d}: '𖨭'

BamumLetterPhaseDashFLoq

\u{16a2e}: '𖨮'

BamumLetterPhaseDashFKo

\u{16a2f}: '𖨯'

BamumLetterPhaseDashFMben

\u{16a30}: '𖨰'

BamumLetterPhaseDashFRen

\u{16a31}: '𖨱'

BamumLetterPhaseDashFMa

\u{16a32}: '𖨲'

BamumLetterPhaseDashFMo

\u{16a33}: '𖨳'

BamumLetterPhaseDashFMbaa

\u{16a34}: '𖨴'

BamumLetterPhaseDashFTet

\u{16a35}: '𖨵'

BamumLetterPhaseDashFKpa

\u{16a36}: '𖨶'

BamumLetterPhaseDashFSamba

\u{16a37}: '𖨷'

BamumLetterPhaseDashFVueq

\u{16a38}: '𖨸'

Methods

impl BamumSupplement

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case