Enum unicode_types::bopomofo::Bopomofo

pub enum Bopomofo {
    LetterB,
    LetterP,
    LetterM,
    LetterF,
    LetterD,
    LetterT,
    LetterN,
    LetterL,
    LetterG,
    LetterK,
    LetterH,
    LetterJ,
    LetterQ,
    LetterX,
    LetterZh,
    LetterCh,
    LetterSh,
    LetterR,
    LetterZ,
    LetterC,
    LetterS,
    LetterA,
    LetterO,
    LetterE,
    LetterEh,
    LetterAi,
    LetterEi,
    LetterAu,
    LetterOu,
    LetterAn,
    LetterEn,
    LetterAng,
    LetterEng,
    LetterEr,
    LetterI,
    LetterU,
    LetterIu,
    LetterV,
    LetterNg,
    LetterGn,
    LetterIh,
    LetterOWithDotAbove,
}

An enum to represent all characters in the Bopomofo block.

Variants
LetterB

\u{3105}: 'ㄅ'

LetterP

\u{3106}: 'ㄆ'

LetterM

\u{3107}: 'ㄇ'

LetterF

\u{3108}: 'ㄈ'

LetterD

\u{3109}: 'ㄉ'

LetterT

\u{310a}: 'ㄊ'

LetterN

\u{310b}: 'ㄋ'

LetterL

\u{310c}: 'ㄌ'

LetterG

\u{310d}: 'ㄍ'

LetterK

\u{310e}: 'ㄎ'

LetterH

\u{310f}: 'ㄏ'

LetterJ

\u{3110}: 'ㄐ'

LetterQ

\u{3111}: 'ㄑ'

LetterX

\u{3112}: 'ㄒ'

LetterZh

\u{3113}: 'ㄓ'

LetterCh

\u{3114}: 'ㄔ'

LetterSh

\u{3115}: 'ㄕ'

LetterR

\u{3116}: 'ㄖ'

LetterZ

\u{3117}: 'ㄗ'

LetterC

\u{3118}: 'ㄘ'

LetterS

\u{3119}: 'ㄙ'

LetterA

\u{311a}: 'ㄚ'

LetterO

\u{311b}: 'ㄛ'

LetterE

\u{311c}: 'ㄜ'

LetterEh

\u{311d}: 'ㄝ'

LetterAi

\u{311e}: 'ㄞ'

LetterEi

\u{311f}: 'ㄟ'

LetterAu

\u{3120}: 'ㄠ'

LetterOu

\u{3121}: 'ㄡ'

LetterAn

\u{3122}: 'ㄢ'

LetterEn

\u{3123}: 'ㄣ'

LetterAng

\u{3124}: 'ㄤ'

LetterEng

\u{3125}: 'ㄥ'

LetterEr

\u{3126}: 'ㄦ'

LetterI

\u{3127}: 'ㄧ'

LetterU

\u{3128}: 'ㄨ'

LetterIu

\u{3129}: 'ㄩ'

LetterV

\u{312a}: 'ㄪ'

LetterNg

\u{312b}: 'ㄫ'

LetterGn

\u{312c}: 'ㄬ'

LetterIh

\u{312d}: 'ㄭ'

LetterOWithDotAbove

\u{312e}: 'ㄮ'

Methods

impl Bopomofo

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case