Enum unicode_types::braille_patterns::BraillePatterns

pub enum BraillePatterns {
    BraillePatternBlank,
    BraillePatternDotsDash1,
    BraillePatternDotsDash2,
    BraillePatternDotsDash12,
    BraillePatternDotsDash3,
    BraillePatternDotsDash13,
    BraillePatternDotsDash23,
    BraillePatternDotsDash123,
    BraillePatternDotsDash4,
    BraillePatternDotsDash14,
    BraillePatternDotsDash24,
    BraillePatternDotsDash124,
    BraillePatternDotsDash34,
    BraillePatternDotsDash134,
    BraillePatternDotsDash234,
    BraillePatternDotsDash1234,
    BraillePatternDotsDash5,
    BraillePatternDotsDash15,
    BraillePatternDotsDash25,
    BraillePatternDotsDash125,
    BraillePatternDotsDash35,
    BraillePatternDotsDash135,
    BraillePatternDotsDash235,
    BraillePatternDotsDash1235,
    BraillePatternDotsDash45,
    BraillePatternDotsDash145,
    BraillePatternDotsDash245,
    BraillePatternDotsDash1245,
    BraillePatternDotsDash345,
    BraillePatternDotsDash1345,
    BraillePatternDotsDash2345,
    BraillePatternDotsDash12345,
    BraillePatternDotsDash6,
    BraillePatternDotsDash16,
    BraillePatternDotsDash26,
    BraillePatternDotsDash126,
    BraillePatternDotsDash36,
    BraillePatternDotsDash136,
    BraillePatternDotsDash236,
    BraillePatternDotsDash1236,
    BraillePatternDotsDash46,
    BraillePatternDotsDash146,
    BraillePatternDotsDash246,
    BraillePatternDotsDash1246,
    BraillePatternDotsDash346,
    BraillePatternDotsDash1346,
    BraillePatternDotsDash2346,
    BraillePatternDotsDash12346,
    BraillePatternDotsDash56,
    BraillePatternDotsDash156,
    BraillePatternDotsDash256,
    BraillePatternDotsDash1256,
    BraillePatternDotsDash356,
    BraillePatternDotsDash1356,
    BraillePatternDotsDash2356,
    BraillePatternDotsDash12356,
    BraillePatternDotsDash456,
    BraillePatternDotsDash1456,
    BraillePatternDotsDash2456,
    BraillePatternDotsDash12456,
    BraillePatternDotsDash3456,
    BraillePatternDotsDash13456,
    BraillePatternDotsDash23456,
    BraillePatternDotsDash123456,
    BraillePatternDotsDash7,
    BraillePatternDotsDash17,
    BraillePatternDotsDash27,
    BraillePatternDotsDash127,
    BraillePatternDotsDash37,
    BraillePatternDotsDash137,
    BraillePatternDotsDash237,
    BraillePatternDotsDash1237,
    BraillePatternDotsDash47,
    BraillePatternDotsDash147,
    BraillePatternDotsDash247,
    BraillePatternDotsDash1247,
    BraillePatternDotsDash347,
    BraillePatternDotsDash1347,
    BraillePatternDotsDash2347,
    BraillePatternDotsDash12347,
    BraillePatternDotsDash57,
    BraillePatternDotsDash157,
    BraillePatternDotsDash257,
    BraillePatternDotsDash1257,
    BraillePatternDotsDash357,
    BraillePatternDotsDash1357,
    BraillePatternDotsDash2357,
    BraillePatternDotsDash12357,
    BraillePatternDotsDash457,
    BraillePatternDotsDash1457,
    BraillePatternDotsDash2457,
    BraillePatternDotsDash12457,
    BraillePatternDotsDash3457,
    BraillePatternDotsDash13457,
    BraillePatternDotsDash23457,
    BraillePatternDotsDash123457,
    BraillePatternDotsDash67,
    BraillePatternDotsDash167,
    BraillePatternDotsDash267,
    BraillePatternDotsDash1267,
    BraillePatternDotsDash367,
    BraillePatternDotsDash1367,
    BraillePatternDotsDash2367,
    BraillePatternDotsDash12367,
    BraillePatternDotsDash467,
    BraillePatternDotsDash1467,
    BraillePatternDotsDash2467,
    BraillePatternDotsDash12467,
    BraillePatternDotsDash3467,
    BraillePatternDotsDash13467,
    BraillePatternDotsDash23467,
    BraillePatternDotsDash123467,
    BraillePatternDotsDash567,
    BraillePatternDotsDash1567,
    BraillePatternDotsDash2567,
    BraillePatternDotsDash12567,
    BraillePatternDotsDash3567,
    BraillePatternDotsDash13567,
    BraillePatternDotsDash23567,
    BraillePatternDotsDash123567,
    BraillePatternDotsDash4567,
    BraillePatternDotsDash14567,
    BraillePatternDotsDash24567,
    BraillePatternDotsDash124567,
    BraillePatternDotsDash34567,
    BraillePatternDotsDash134567,
    BraillePatternDotsDash234567,
    BraillePatternDotsDash1234567,
    BraillePatternDotsDash8,
    BraillePatternDotsDash18,
    BraillePatternDotsDash28,
    BraillePatternDotsDash128,
    BraillePatternDotsDash38,
    BraillePatternDotsDash138,
    BraillePatternDotsDash238,
    BraillePatternDotsDash1238,
    BraillePatternDotsDash48,
    BraillePatternDotsDash148,
    BraillePatternDotsDash248,
    BraillePatternDotsDash1248,
    BraillePatternDotsDash348,
    BraillePatternDotsDash1348,
    BraillePatternDotsDash2348,
    BraillePatternDotsDash12348,
    BraillePatternDotsDash58,
    BraillePatternDotsDash158,
    BraillePatternDotsDash258,
    BraillePatternDotsDash1258,
    BraillePatternDotsDash358,
    BraillePatternDotsDash1358,
    BraillePatternDotsDash2358,
    BraillePatternDotsDash12358,
    BraillePatternDotsDash458,
    BraillePatternDotsDash1458,
    BraillePatternDotsDash2458,
    BraillePatternDotsDash12458,
    BraillePatternDotsDash3458,
    BraillePatternDotsDash13458,
    BraillePatternDotsDash23458,
    BraillePatternDotsDash123458,
    BraillePatternDotsDash68,
    BraillePatternDotsDash168,
    BraillePatternDotsDash268,
    BraillePatternDotsDash1268,
    BraillePatternDotsDash368,
    BraillePatternDotsDash1368,
    BraillePatternDotsDash2368,
    BraillePatternDotsDash12368,
    BraillePatternDotsDash468,
    BraillePatternDotsDash1468,
    BraillePatternDotsDash2468,
    BraillePatternDotsDash12468,
    BraillePatternDotsDash3468,
    BraillePatternDotsDash13468,
    BraillePatternDotsDash23468,
    BraillePatternDotsDash123468,
    BraillePatternDotsDash568,
    BraillePatternDotsDash1568,
    BraillePatternDotsDash2568,
    BraillePatternDotsDash12568,
    BraillePatternDotsDash3568,
    BraillePatternDotsDash13568,
    BraillePatternDotsDash23568,
    BraillePatternDotsDash123568,
    BraillePatternDotsDash4568,
    BraillePatternDotsDash14568,
    BraillePatternDotsDash24568,
    BraillePatternDotsDash124568,
    BraillePatternDotsDash34568,
    BraillePatternDotsDash134568,
    BraillePatternDotsDash234568,
    BraillePatternDotsDash1234568,
    BraillePatternDotsDash78,
    BraillePatternDotsDash178,
    BraillePatternDotsDash278,
    BraillePatternDotsDash1278,
    BraillePatternDotsDash378,
    BraillePatternDotsDash1378,
    BraillePatternDotsDash2378,
    BraillePatternDotsDash12378,
    BraillePatternDotsDash478,
    BraillePatternDotsDash1478,
    BraillePatternDotsDash2478,
    BraillePatternDotsDash12478,
    BraillePatternDotsDash3478,
    BraillePatternDotsDash13478,
    BraillePatternDotsDash23478,
    BraillePatternDotsDash123478,
    BraillePatternDotsDash578,
    BraillePatternDotsDash1578,
    BraillePatternDotsDash2578,
    BraillePatternDotsDash12578,
    BraillePatternDotsDash3578,
    BraillePatternDotsDash13578,
    BraillePatternDotsDash23578,
    BraillePatternDotsDash123578,
    BraillePatternDotsDash4578,
    BraillePatternDotsDash14578,
    BraillePatternDotsDash24578,
    BraillePatternDotsDash124578,
    BraillePatternDotsDash34578,
    BraillePatternDotsDash134578,
    BraillePatternDotsDash234578,
    BraillePatternDotsDash1234578,
    BraillePatternDotsDash678,
    BraillePatternDotsDash1678,
    BraillePatternDotsDash2678,
    BraillePatternDotsDash12678,
    BraillePatternDotsDash3678,
    BraillePatternDotsDash13678,
    BraillePatternDotsDash23678,
    BraillePatternDotsDash123678,
    BraillePatternDotsDash4678,
    BraillePatternDotsDash14678,
    BraillePatternDotsDash24678,
    BraillePatternDotsDash124678,
    BraillePatternDotsDash34678,
    BraillePatternDotsDash134678,
    BraillePatternDotsDash234678,
    BraillePatternDotsDash1234678,
    BraillePatternDotsDash5678,
    BraillePatternDotsDash15678,
    BraillePatternDotsDash25678,
    BraillePatternDotsDash125678,
    BraillePatternDotsDash35678,
    BraillePatternDotsDash135678,
    BraillePatternDotsDash235678,
    BraillePatternDotsDash1235678,
    BraillePatternDotsDash45678,
    BraillePatternDotsDash145678,
    BraillePatternDotsDash245678,
    BraillePatternDotsDash1245678,
    BraillePatternDotsDash345678,
    BraillePatternDotsDash1345678,
    BraillePatternDotsDash2345678,
}

An enum to represent all characters in the BraillePatterns block.

Variants
BraillePatternBlank

\u{2800}: '⠀'

BraillePatternDotsDash1

\u{2801}: '⠁'

BraillePatternDotsDash2

\u{2802}: '⠂'

BraillePatternDotsDash12

\u{2803}: '⠃'

BraillePatternDotsDash3

\u{2804}: '⠄'

BraillePatternDotsDash13

\u{2805}: '⠅'

BraillePatternDotsDash23

\u{2806}: '⠆'

BraillePatternDotsDash123

\u{2807}: '⠇'

BraillePatternDotsDash4

\u{2808}: '⠈'

BraillePatternDotsDash14

\u{2809}: '⠉'

BraillePatternDotsDash24

\u{280a}: '⠊'

BraillePatternDotsDash124

\u{280b}: '⠋'

BraillePatternDotsDash34

\u{280c}: '⠌'

BraillePatternDotsDash134

\u{280d}: '⠍'

BraillePatternDotsDash234

\u{280e}: '⠎'

BraillePatternDotsDash1234

\u{280f}: '⠏'

BraillePatternDotsDash5

\u{2810}: '⠐'

BraillePatternDotsDash15

\u{2811}: '⠑'

BraillePatternDotsDash25

\u{2812}: '⠒'

BraillePatternDotsDash125

\u{2813}: '⠓'

BraillePatternDotsDash35

\u{2814}: '⠔'

BraillePatternDotsDash135

\u{2815}: '⠕'

BraillePatternDotsDash235

\u{2816}: '⠖'

BraillePatternDotsDash1235

\u{2817}: '⠗'

BraillePatternDotsDash45

\u{2818}: '⠘'

BraillePatternDotsDash145

\u{2819}: '⠙'

BraillePatternDotsDash245

\u{281a}: '⠚'

BraillePatternDotsDash1245

\u{281b}: '⠛'

BraillePatternDotsDash345

\u{281c}: '⠜'

BraillePatternDotsDash1345

\u{281d}: '⠝'

BraillePatternDotsDash2345

\u{281e}: '⠞'

BraillePatternDotsDash12345

\u{281f}: '⠟'

BraillePatternDotsDash6

\u{2820}: '⠠'

BraillePatternDotsDash16

\u{2821}: '⠡'

BraillePatternDotsDash26

\u{2822}: '⠢'

BraillePatternDotsDash126

\u{2823}: '⠣'

BraillePatternDotsDash36

\u{2824}: '⠤'

BraillePatternDotsDash136

\u{2825}: '⠥'

BraillePatternDotsDash236

\u{2826}: '⠦'

BraillePatternDotsDash1236

\u{2827}: '⠧'

BraillePatternDotsDash46

\u{2828}: '⠨'

BraillePatternDotsDash146

\u{2829}: '⠩'

BraillePatternDotsDash246

\u{282a}: '⠪'

BraillePatternDotsDash1246

\u{282b}: '⠫'

BraillePatternDotsDash346

\u{282c}: '⠬'

BraillePatternDotsDash1346

\u{282d}: '⠭'

BraillePatternDotsDash2346

\u{282e}: '⠮'

BraillePatternDotsDash12346

\u{282f}: '⠯'

BraillePatternDotsDash56

\u{2830}: '⠰'

BraillePatternDotsDash156

\u{2831}: '⠱'

BraillePatternDotsDash256

\u{2832}: '⠲'

BraillePatternDotsDash1256

\u{2833}: '⠳'

BraillePatternDotsDash356

\u{2834}: '⠴'

BraillePatternDotsDash1356

\u{2835}: '⠵'

BraillePatternDotsDash2356

\u{2836}: '⠶'

BraillePatternDotsDash12356

\u{2837}: '⠷'

BraillePatternDotsDash456

\u{2838}: '⠸'

BraillePatternDotsDash1456

\u{2839}: '⠹'

BraillePatternDotsDash2456

\u{283a}: '⠺'

BraillePatternDotsDash12456

\u{283b}: '⠻'

BraillePatternDotsDash3456

\u{283c}: '⠼'

BraillePatternDotsDash13456

\u{283d}: '⠽'

BraillePatternDotsDash23456

\u{283e}: '⠾'

BraillePatternDotsDash123456

\u{283f}: '⠿'

BraillePatternDotsDash7

\u{2840}: '⡀'

BraillePatternDotsDash17

\u{2841}: '⡁'

BraillePatternDotsDash27

\u{2842}: '⡂'

BraillePatternDotsDash127

\u{2843}: '⡃'

BraillePatternDotsDash37

\u{2844}: '⡄'

BraillePatternDotsDash137

\u{2845}: '⡅'

BraillePatternDotsDash237

\u{2846}: '⡆'

BraillePatternDotsDash1237

\u{2847}: '⡇'

BraillePatternDotsDash47

\u{2848}: '⡈'

BraillePatternDotsDash147

\u{2849}: '⡉'

BraillePatternDotsDash247

\u{284a}: '⡊'

BraillePatternDotsDash1247

\u{284b}: '⡋'

BraillePatternDotsDash347

\u{284c}: '⡌'

BraillePatternDotsDash1347

\u{284d}: '⡍'

BraillePatternDotsDash2347

\u{284e}: '⡎'

BraillePatternDotsDash12347

\u{284f}: '⡏'

BraillePatternDotsDash57

\u{2850}: '⡐'

BraillePatternDotsDash157

\u{2851}: '⡑'

BraillePatternDotsDash257

\u{2852}: '⡒'

BraillePatternDotsDash1257

\u{2853}: '⡓'

BraillePatternDotsDash357

\u{2854}: '⡔'

BraillePatternDotsDash1357

\u{2855}: '⡕'

BraillePatternDotsDash2357

\u{2856}: '⡖'

BraillePatternDotsDash12357

\u{2857}: '⡗'

BraillePatternDotsDash457

\u{2858}: '⡘'

BraillePatternDotsDash1457

\u{2859}: '⡙'

BraillePatternDotsDash2457

\u{285a}: '⡚'

BraillePatternDotsDash12457

\u{285b}: '⡛'

BraillePatternDotsDash3457

\u{285c}: '⡜'

BraillePatternDotsDash13457

\u{285d}: '⡝'

BraillePatternDotsDash23457

\u{285e}: '⡞'

BraillePatternDotsDash123457

\u{285f}: '⡟'

BraillePatternDotsDash67

\u{2860}: '⡠'

BraillePatternDotsDash167

\u{2861}: '⡡'

BraillePatternDotsDash267

\u{2862}: '⡢'

BraillePatternDotsDash1267

\u{2863}: '⡣'

BraillePatternDotsDash367

\u{2864}: '⡤'

BraillePatternDotsDash1367

\u{2865}: '⡥'

BraillePatternDotsDash2367

\u{2866}: '⡦'

BraillePatternDotsDash12367

\u{2867}: '⡧'

BraillePatternDotsDash467

\u{2868}: '⡨'

BraillePatternDotsDash1467

\u{2869}: '⡩'

BraillePatternDotsDash2467

\u{286a}: '⡪'

BraillePatternDotsDash12467

\u{286b}: '⡫'

BraillePatternDotsDash3467

\u{286c}: '⡬'

BraillePatternDotsDash13467

\u{286d}: '⡭'

BraillePatternDotsDash23467

\u{286e}: '⡮'

BraillePatternDotsDash123467

\u{286f}: '⡯'

BraillePatternDotsDash567

\u{2870}: '⡰'

BraillePatternDotsDash1567

\u{2871}: '⡱'

BraillePatternDotsDash2567

\u{2872}: '⡲'

BraillePatternDotsDash12567

\u{2873}: '⡳'

BraillePatternDotsDash3567

\u{2874}: '⡴'

BraillePatternDotsDash13567

\u{2875}: '⡵'

BraillePatternDotsDash23567

\u{2876}: '⡶'

BraillePatternDotsDash123567

\u{2877}: '⡷'

BraillePatternDotsDash4567

\u{2878}: '⡸'

BraillePatternDotsDash14567

\u{2879}: '⡹'

BraillePatternDotsDash24567

\u{287a}: '⡺'

BraillePatternDotsDash124567

\u{287b}: '⡻'

BraillePatternDotsDash34567

\u{287c}: '⡼'

BraillePatternDotsDash134567

\u{287d}: '⡽'

BraillePatternDotsDash234567

\u{287e}: '⡾'

BraillePatternDotsDash1234567

\u{287f}: '⡿'

BraillePatternDotsDash8

\u{2880}: '⢀'

BraillePatternDotsDash18

\u{2881}: '⢁'

BraillePatternDotsDash28

\u{2882}: '⢂'

BraillePatternDotsDash128

\u{2883}: '⢃'

BraillePatternDotsDash38

\u{2884}: '⢄'

BraillePatternDotsDash138

\u{2885}: '⢅'

BraillePatternDotsDash238

\u{2886}: '⢆'

BraillePatternDotsDash1238

\u{2887}: '⢇'

BraillePatternDotsDash48

\u{2888}: '⢈'

BraillePatternDotsDash148

\u{2889}: '⢉'

BraillePatternDotsDash248

\u{288a}: '⢊'

BraillePatternDotsDash1248

\u{288b}: '⢋'

BraillePatternDotsDash348

\u{288c}: '⢌'

BraillePatternDotsDash1348

\u{288d}: '⢍'

BraillePatternDotsDash2348

\u{288e}: '⢎'

BraillePatternDotsDash12348

\u{288f}: '⢏'

BraillePatternDotsDash58

\u{2890}: '⢐'

BraillePatternDotsDash158

\u{2891}: '⢑'

BraillePatternDotsDash258

\u{2892}: '⢒'

BraillePatternDotsDash1258

\u{2893}: '⢓'

BraillePatternDotsDash358

\u{2894}: '⢔'

BraillePatternDotsDash1358

\u{2895}: '⢕'

BraillePatternDotsDash2358

\u{2896}: '⢖'

BraillePatternDotsDash12358

\u{2897}: '⢗'

BraillePatternDotsDash458

\u{2898}: '⢘'

BraillePatternDotsDash1458

\u{2899}: '⢙'

BraillePatternDotsDash2458

\u{289a}: '⢚'

BraillePatternDotsDash12458

\u{289b}: '⢛'

BraillePatternDotsDash3458

\u{289c}: '⢜'

BraillePatternDotsDash13458

\u{289d}: '⢝'

BraillePatternDotsDash23458

\u{289e}: '⢞'

BraillePatternDotsDash123458

\u{289f}: '⢟'

BraillePatternDotsDash68

\u{28a0}: '⢠'

BraillePatternDotsDash168

\u{28a1}: '⢡'

BraillePatternDotsDash268

\u{28a2}: '⢢'

BraillePatternDotsDash1268

\u{28a3}: '⢣'

BraillePatternDotsDash368

\u{28a4}: '⢤'

BraillePatternDotsDash1368

\u{28a5}: '⢥'

BraillePatternDotsDash2368

\u{28a6}: '⢦'

BraillePatternDotsDash12368

\u{28a7}: '⢧'

BraillePatternDotsDash468

\u{28a8}: '⢨'

BraillePatternDotsDash1468

\u{28a9}: '⢩'

BraillePatternDotsDash2468

\u{28aa}: '⢪'

BraillePatternDotsDash12468

\u{28ab}: '⢫'

BraillePatternDotsDash3468

\u{28ac}: '⢬'

BraillePatternDotsDash13468

\u{28ad}: '⢭'

BraillePatternDotsDash23468

\u{28ae}: '⢮'

BraillePatternDotsDash123468

\u{28af}: '⢯'

BraillePatternDotsDash568

\u{28b0}: '⢰'

BraillePatternDotsDash1568

\u{28b1}: '⢱'

BraillePatternDotsDash2568

\u{28b2}: '⢲'

BraillePatternDotsDash12568

\u{28b3}: '⢳'

BraillePatternDotsDash3568

\u{28b4}: '⢴'

BraillePatternDotsDash13568

\u{28b5}: '⢵'

BraillePatternDotsDash23568

\u{28b6}: '⢶'

BraillePatternDotsDash123568

\u{28b7}: '⢷'

BraillePatternDotsDash4568

\u{28b8}: '⢸'

BraillePatternDotsDash14568

\u{28b9}: '⢹'

BraillePatternDotsDash24568

\u{28ba}: '⢺'

BraillePatternDotsDash124568

\u{28bb}: '⢻'

BraillePatternDotsDash34568

\u{28bc}: '⢼'

BraillePatternDotsDash134568

\u{28bd}: '⢽'

BraillePatternDotsDash234568

\u{28be}: '⢾'

BraillePatternDotsDash1234568

\u{28bf}: '⢿'

BraillePatternDotsDash78

\u{28c0}: '⣀'

BraillePatternDotsDash178

\u{28c1}: '⣁'

BraillePatternDotsDash278

\u{28c2}: '⣂'

BraillePatternDotsDash1278

\u{28c3}: '⣃'

BraillePatternDotsDash378

\u{28c4}: '⣄'

BraillePatternDotsDash1378

\u{28c5}: '⣅'

BraillePatternDotsDash2378

\u{28c6}: '⣆'

BraillePatternDotsDash12378

\u{28c7}: '⣇'

BraillePatternDotsDash478

\u{28c8}: '⣈'

BraillePatternDotsDash1478

\u{28c9}: '⣉'

BraillePatternDotsDash2478

\u{28ca}: '⣊'

BraillePatternDotsDash12478

\u{28cb}: '⣋'

BraillePatternDotsDash3478

\u{28cc}: '⣌'

BraillePatternDotsDash13478

\u{28cd}: '⣍'

BraillePatternDotsDash23478

\u{28ce}: '⣎'

BraillePatternDotsDash123478

\u{28cf}: '⣏'

BraillePatternDotsDash578

\u{28d0}: '⣐'

BraillePatternDotsDash1578

\u{28d1}: '⣑'

BraillePatternDotsDash2578

\u{28d2}: '⣒'

BraillePatternDotsDash12578

\u{28d3}: '⣓'

BraillePatternDotsDash3578

\u{28d4}: '⣔'

BraillePatternDotsDash13578

\u{28d5}: '⣕'

BraillePatternDotsDash23578

\u{28d6}: '⣖'

BraillePatternDotsDash123578

\u{28d7}: '⣗'

BraillePatternDotsDash4578

\u{28d8}: '⣘'

BraillePatternDotsDash14578

\u{28d9}: '⣙'

BraillePatternDotsDash24578

\u{28da}: '⣚'

BraillePatternDotsDash124578

\u{28db}: '⣛'

BraillePatternDotsDash34578

\u{28dc}: '⣜'

BraillePatternDotsDash134578

\u{28dd}: '⣝'

BraillePatternDotsDash234578

\u{28de}: '⣞'

BraillePatternDotsDash1234578

\u{28df}: '⣟'

BraillePatternDotsDash678

\u{28e0}: '⣠'

BraillePatternDotsDash1678

\u{28e1}: '⣡'

BraillePatternDotsDash2678

\u{28e2}: '⣢'

BraillePatternDotsDash12678

\u{28e3}: '⣣'

BraillePatternDotsDash3678

\u{28e4}: '⣤'

BraillePatternDotsDash13678

\u{28e5}: '⣥'

BraillePatternDotsDash23678

\u{28e6}: '⣦'

BraillePatternDotsDash123678

\u{28e7}: '⣧'

BraillePatternDotsDash4678

\u{28e8}: '⣨'

BraillePatternDotsDash14678

\u{28e9}: '⣩'

BraillePatternDotsDash24678

\u{28ea}: '⣪'

BraillePatternDotsDash124678

\u{28eb}: '⣫'

BraillePatternDotsDash34678

\u{28ec}: '⣬'

BraillePatternDotsDash134678

\u{28ed}: '⣭'

BraillePatternDotsDash234678

\u{28ee}: '⣮'

BraillePatternDotsDash1234678

\u{28ef}: '⣯'

BraillePatternDotsDash5678

\u{28f0}: '⣰'

BraillePatternDotsDash15678

\u{28f1}: '⣱'

BraillePatternDotsDash25678

\u{28f2}: '⣲'

BraillePatternDotsDash125678

\u{28f3}: '⣳'

BraillePatternDotsDash35678

\u{28f4}: '⣴'

BraillePatternDotsDash135678

\u{28f5}: '⣵'

BraillePatternDotsDash235678

\u{28f6}: '⣶'

BraillePatternDotsDash1235678

\u{28f7}: '⣷'

BraillePatternDotsDash45678

\u{28f8}: '⣸'

BraillePatternDotsDash145678

\u{28f9}: '⣹'

BraillePatternDotsDash245678

\u{28fa}: '⣺'

BraillePatternDotsDash1245678

\u{28fb}: '⣻'

BraillePatternDotsDash345678

\u{28fc}: '⣼'

BraillePatternDotsDash1345678

\u{28fd}: '⣽'

BraillePatternDotsDash2345678

\u{28fe}: '⣾'

Methods

impl BraillePatterns

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case