Enum unicode_types::byzantine_musical_symbols::ByzantineMusicalSymbols

pub enum ByzantineMusicalSymbols {
    ByzantineMusicalSymbolPsili,
    ByzantineMusicalSymbolDaseia,
    ByzantineMusicalSymbolPerispomeni,
    ByzantineMusicalSymbolOxeiaEkfonitikon,
    ByzantineMusicalSymbolOxeiaDipli,
    ByzantineMusicalSymbolVareiaEkfonitikon,
    ByzantineMusicalSymbolVareiaDipli,
    ByzantineMusicalSymbolKathisti,
    ByzantineMusicalSymbolSyrmatiki,
    ByzantineMusicalSymbolParaklitiki,
    ByzantineMusicalSymbolYpokrisis,
    ByzantineMusicalSymbolYpokrisisDipli,
    ByzantineMusicalSymbolKremasti,
    ByzantineMusicalSymbolApesoEkfonitikon,
    ByzantineMusicalSymbolExoEkfonitikon,
    ByzantineMusicalSymbolTeleia,
    ByzantineMusicalSymbolKentimata,
    ByzantineMusicalSymbolApostrofos,
    ByzantineMusicalSymbolApostrofosDipli,
    ByzantineMusicalSymbolSynevma,
    ByzantineMusicalSymbolThita,
    ByzantineMusicalSymbolOligonArchaion,
    ByzantineMusicalSymbolGorgonArchaion,
    ByzantineMusicalSymbolPsilon,
    ByzantineMusicalSymbolChamilon,
    ByzantineMusicalSymbolVathy,
    ByzantineMusicalSymbolIsonArchaion,
    ByzantineMusicalSymbolKentimaArchaion,
    ByzantineMusicalSymbolKentimataArchaion,
    ByzantineMusicalSymbolSaximata,
    ByzantineMusicalSymbolParichon,
    ByzantineMusicalSymbolStavrosApodexia,
    ByzantineMusicalSymbolOxeiaiArchaion,
    ByzantineMusicalSymbolVareiaiArchaion,
    ByzantineMusicalSymbolApodermaArchaion,
    ByzantineMusicalSymbolApothema,
    ByzantineMusicalSymbolKlasma,
    ByzantineMusicalSymbolRevma,
    ByzantineMusicalSymbolPiasmaArchaion,
    ByzantineMusicalSymbolTinagma,
    ByzantineMusicalSymbolAnatrichisma,
    ByzantineMusicalSymbolSeisma,
    ByzantineMusicalSymbolSynagmaArchaion,
    ByzantineMusicalSymbolSynagmaMetaStavrou,
    ByzantineMusicalSymbolOyranismaArchaion,
    ByzantineMusicalSymbolThema,
    ByzantineMusicalSymbolLemoi,
    ByzantineMusicalSymbolDyo,
    ByzantineMusicalSymbolTria,
    ByzantineMusicalSymbolTessera,
    ByzantineMusicalSymbolKratimata,
    ByzantineMusicalSymbolApesoExoNeo,
    ByzantineMusicalSymbolFthoraArchaion,
    ByzantineMusicalSymbolImifthora,
    ByzantineMusicalSymbolTromikonArchaion,
    ByzantineMusicalSymbolKatavaTromikon,
    ByzantineMusicalSymbolPelaston,
    ByzantineMusicalSymbolPsifiston,
    ByzantineMusicalSymbolKontevma,
    ByzantineMusicalSymbolChorevmaArchaion,
    ByzantineMusicalSymbolRapisma,
    ByzantineMusicalSymbolParakalesmaArchaion,
    ByzantineMusicalSymbolParaklitikiArchaion,
    ByzantineMusicalSymbolIchadin,
    ByzantineMusicalSymbolNana,
    ByzantineMusicalSymbolPetasma,
    ByzantineMusicalSymbolKontevmaAllo,
    ByzantineMusicalSymbolTromikonAllo,
    ByzantineMusicalSymbolStraggismata,
    ByzantineMusicalSymbolGronthismata,
    ByzantineMusicalSymbolIsonNeo,
    ByzantineMusicalSymbolOligonNeo,
    ByzantineMusicalSymbolOxeiaNeo,
    ByzantineMusicalSymbolPetasti,
    ByzantineMusicalSymbolKoufisma,
    ByzantineMusicalSymbolPetastokoufisma,
    ByzantineMusicalSymbolKratimokoufisma,
    ByzantineMusicalSymbolPelastonNeo,
    ByzantineMusicalSymbolKentimataNeoAno,
    ByzantineMusicalSymbolKentimaNeoAno,
    ByzantineMusicalSymbolYpsili,
    ByzantineMusicalSymbolApostrofosNeo,
    ByzantineMusicalSymbolApostrofoiSyndesmosNeo,
    ByzantineMusicalSymbolYporroi,
    ByzantineMusicalSymbolKratimoyporroon,
    ByzantineMusicalSymbolElafron,
    ByzantineMusicalSymbolChamili,
    ByzantineMusicalSymbolMikronIson,
    ByzantineMusicalSymbolVareiaNeo,
    ByzantineMusicalSymbolPiasmaNeo,
    ByzantineMusicalSymbolPsifistonNeo,
    ByzantineMusicalSymbolOmalon,
    ByzantineMusicalSymbolAntikenoma,
    ByzantineMusicalSymbolLygisma,
    ByzantineMusicalSymbolParaklitikiNeo,
    ByzantineMusicalSymbolParakalesmaNeo,
    ByzantineMusicalSymbolEteronParakalesma,
    ByzantineMusicalSymbolKylisma,
    ByzantineMusicalSymbolAntikenokylisma,
    ByzantineMusicalSymbolTromikonNeo,
    ByzantineMusicalSymbolEkstrepton,
    ByzantineMusicalSymbolSynagmaNeo,
    ByzantineMusicalSymbolSyrma,
    ByzantineMusicalSymbolChorevmaNeo,
    ByzantineMusicalSymbolEpegerma,
    ByzantineMusicalSymbolSeismaNeo,
    ByzantineMusicalSymbolXironKlasma,
    ByzantineMusicalSymbolTromikopsifiston,
    ByzantineMusicalSymbolPsifistolygisma,
    ByzantineMusicalSymbolTromikolygisma,
    ByzantineMusicalSymbolTromikoparakalesma,
    ByzantineMusicalSymbolPsifistoparakalesma,
    ByzantineMusicalSymbolTromikosynagma,
    ByzantineMusicalSymbolPsifistosynagma,
    ByzantineMusicalSymbolGorgosyntheton,
    ByzantineMusicalSymbolArgosyntheton,
    ByzantineMusicalSymbolEteronArgosyntheton,
    ByzantineMusicalSymbolOyranismaNeo,
    ByzantineMusicalSymbolThematismosEso,
    ByzantineMusicalSymbolThematismosExo,
    ByzantineMusicalSymbolThemaAploun,
    ByzantineMusicalSymbolThesKaiApothes,
    ByzantineMusicalSymbolKatavasma,
    ByzantineMusicalSymbolEndofonon,
    ByzantineMusicalSymbolYfenKato,
    ByzantineMusicalSymbolYfenAno,
    ByzantineMusicalSymbolStavros,
    ByzantineMusicalSymbolKlasmaAno,
    ByzantineMusicalSymbolDipliArchaion,
    ByzantineMusicalSymbolKratimaArchaion,
    ByzantineMusicalSymbolKratimaAllo,
    ByzantineMusicalSymbolKratimaNeo,
    ByzantineMusicalSymbolApodermaNeo,
    ByzantineMusicalSymbolApli,
    ByzantineMusicalSymbolDipli,
    ByzantineMusicalSymbolTripli,
    ByzantineMusicalSymbolTetrapli,
    ByzantineMusicalSymbolKoronis,
    ByzantineMusicalSymbolLeimmaEnosChronou,
    ByzantineMusicalSymbolLeimmaDyoChronon,
    ByzantineMusicalSymbolLeimmaTrionChronon,
    ByzantineMusicalSymbolLeimmaTessaronChronon,
    ByzantineMusicalSymbolLeimmaImiseosChronou,
    ByzantineMusicalSymbolGorgonNeoAno,
    ByzantineMusicalSymbolGorgonParestigmenonAristera,
    ByzantineMusicalSymbolGorgonParestigmenonDexia,
    ByzantineMusicalSymbolDigorgon,
    ByzantineMusicalSymbolDigorgonParestigmenonAristeraKato,
    ByzantineMusicalSymbolDigorgonParestigmenonAristeraAno,
    ByzantineMusicalSymbolDigorgonParestigmenonDexia,
    ByzantineMusicalSymbolTrigorgon,
    ByzantineMusicalSymbolArgon,
    ByzantineMusicalSymbolImidiargon,
    ByzantineMusicalSymbolDiargon,
    ByzantineMusicalSymbolAgogiPoliArgi,
    ByzantineMusicalSymbolAgogiArgoteri,
    ByzantineMusicalSymbolAgogiArgi,
    ByzantineMusicalSymbolAgogiMetria,
    ByzantineMusicalSymbolAgogiMesi,
    ByzantineMusicalSymbolAgogiGorgi,
    ByzantineMusicalSymbolAgogiGorgoteri,
    ByzantineMusicalSymbolAgogiPoliGorgi,
    ByzantineMusicalSymbolMartyriaProtosIchos,
    ByzantineMusicalSymbolMartyriaAlliProtosIchos,
    ByzantineMusicalSymbolMartyriaDeyterosIchos,
    ByzantineMusicalSymbolMartyriaAlliDeyterosIchos,
    ByzantineMusicalSymbolMartyriaTritosIchos,
    ByzantineMusicalSymbolMartyriaTrifonias,
    ByzantineMusicalSymbolMartyriaTetartosIchos,
    ByzantineMusicalSymbolMartyriaTetartosLegetosIchos,
    ByzantineMusicalSymbolMartyriaLegetosIchos,
    ByzantineMusicalSymbolMartyriaPlagiosIchos,
    ByzantineMusicalSymbolIsakiaTelousIchimatos,
    ByzantineMusicalSymbolApostrofoiTelousIchimatos,
    ByzantineMusicalSymbolFanerosisTetrafonias,
    ByzantineMusicalSymbolFanerosisMonofonias,
    ByzantineMusicalSymbolFanerosisDifonias,
    ByzantineMusicalSymbolMartyriaVarysIchos,
    ByzantineMusicalSymbolMartyriaProtovarysIchos,
    ByzantineMusicalSymbolMartyriaPlagiosTetartosIchos,
    ByzantineMusicalSymbolGorthmikonNAploun,
    ByzantineMusicalSymbolGorthmikonNDiploun,
    ByzantineMusicalSymbolEnarxisKaiFthoraVou,
    ByzantineMusicalSymbolImifonon,
    ByzantineMusicalSymbolImifthoron,
    ByzantineMusicalSymbolFthoraArchaionDeyterouIchou,
    ByzantineMusicalSymbolFthoraDiatonikiPa,
    ByzantineMusicalSymbolFthoraDiatonikiNana,
    ByzantineMusicalSymbolFthoraNaosIchos,
    ByzantineMusicalSymbolFthoraDiatonikiDi,
    ByzantineMusicalSymbolFthoraSklironDiatononDi,
    ByzantineMusicalSymbolFthoraDiatonikiKe,
    ByzantineMusicalSymbolFthoraDiatonikiZo,
    ByzantineMusicalSymbolFthoraDiatonikiNiKato,
    ByzantineMusicalSymbolFthoraDiatonikiNiAno,
    ByzantineMusicalSymbolFthoraMalakonChromaDifonias,
    ByzantineMusicalSymbolFthoraMalakonChromaMonofonias,
    ByzantineMusicalSymbolFhtoraSklironChromaVasis,
    ByzantineMusicalSymbolFthoraSklironChromaSynafi,
    ByzantineMusicalSymbolFthoraNenano,
    ByzantineMusicalSymbolChroaZygos,
    ByzantineMusicalSymbolChroaKliton,
    ByzantineMusicalSymbolChroaSpathi,
    ByzantineMusicalSymbolFthoraIYfesisTetartimorion,
    ByzantineMusicalSymbolFthoraEnarmoniosAntifonia,
    ByzantineMusicalSymbolYfesisTritimorion,
    ByzantineMusicalSymbolDiesisTritimorion,
    ByzantineMusicalSymbolDiesisTetartimorion,
    ByzantineMusicalSymbolDiesisApliDyoDodekata,
    ByzantineMusicalSymbolDiesisMonogrammosTesseraDodekata,
    ByzantineMusicalSymbolDiesisDigrammosExDodekata,
    ByzantineMusicalSymbolDiesisTrigrammosOktoDodekata,
    ByzantineMusicalSymbolYfesisApliDyoDodekata,
    ByzantineMusicalSymbolYfesisMonogrammosTesseraDodekata,
    ByzantineMusicalSymbolYfesisDigrammosExDodekata,
    ByzantineMusicalSymbolYfesisTrigrammosOktoDodekata,
    ByzantineMusicalSymbolGenikiDiesis,
    ByzantineMusicalSymbolGenikiYfesis,
    ByzantineMusicalSymbolDiastoliApliMikri,
    ByzantineMusicalSymbolDiastoliApliMegali,
    ByzantineMusicalSymbolDiastoliDipli,
    ByzantineMusicalSymbolDiastoliTheseos,
    ByzantineMusicalSymbolSimansisTheseos,
    ByzantineMusicalSymbolSimansisTheseosDisimou,
    ByzantineMusicalSymbolSimansisTheseosTrisimou,
    ByzantineMusicalSymbolSimansisTheseosTetrasimou,
    ByzantineMusicalSymbolSimansisArseos,
    ByzantineMusicalSymbolSimansisArseosDisimou,
    ByzantineMusicalSymbolSimansisArseosTrisimou,
    ByzantineMusicalSymbolSimansisArseosTetrasimou,
    ByzantineMusicalSymbolDigrammaGg,
    ByzantineMusicalSymbolDiftoggosOu,
    ByzantineMusicalSymbolStigma,
    ByzantineMusicalSymbolArktikoPa,
    ByzantineMusicalSymbolArktikoVou,
    ByzantineMusicalSymbolArktikoGa,
    ByzantineMusicalSymbolArktikoDi,
    ByzantineMusicalSymbolArktikoKe,
    ByzantineMusicalSymbolArktikoZo,
    ByzantineMusicalSymbolArktikoNi,
    ByzantineMusicalSymbolKentimataNeoMeso,
    ByzantineMusicalSymbolKentimaNeoMeso,
    ByzantineMusicalSymbolKentimataNeoKato,
    ByzantineMusicalSymbolKentimaNeoKato,
    ByzantineMusicalSymbolKlasmaKato,
    ByzantineMusicalSymbolGorgonNeoKato,
}

An enum to represent all characters in the ByzantineMusicalSymbols block.

Variants
ByzantineMusicalSymbolPsili

\u{1d000}: '𝀀'

ByzantineMusicalSymbolDaseia

\u{1d001}: '𝀁'

ByzantineMusicalSymbolPerispomeni

\u{1d002}: '𝀂'

ByzantineMusicalSymbolOxeiaEkfonitikon

\u{1d003}: '𝀃'

ByzantineMusicalSymbolOxeiaDipli

\u{1d004}: '𝀄'

ByzantineMusicalSymbolVareiaEkfonitikon

\u{1d005}: '𝀅'

ByzantineMusicalSymbolVareiaDipli

\u{1d006}: '𝀆'

ByzantineMusicalSymbolKathisti

\u{1d007}: '𝀇'

ByzantineMusicalSymbolSyrmatiki

\u{1d008}: '𝀈'

ByzantineMusicalSymbolParaklitiki

\u{1d009}: '𝀉'

ByzantineMusicalSymbolYpokrisis

\u{1d00a}: '𝀊'

ByzantineMusicalSymbolYpokrisisDipli

\u{1d00b}: '𝀋'

ByzantineMusicalSymbolKremasti

\u{1d00c}: '𝀌'

ByzantineMusicalSymbolApesoEkfonitikon

\u{1d00d}: '𝀍'

ByzantineMusicalSymbolExoEkfonitikon

\u{1d00e}: '𝀎'

ByzantineMusicalSymbolTeleia

\u{1d00f}: '𝀏'

ByzantineMusicalSymbolKentimata

\u{1d010}: '𝀐'

ByzantineMusicalSymbolApostrofos

\u{1d011}: '𝀑'

ByzantineMusicalSymbolApostrofosDipli

\u{1d012}: '𝀒'

ByzantineMusicalSymbolSynevma

\u{1d013}: '𝀓'

ByzantineMusicalSymbolThita

\u{1d014}: '𝀔'

ByzantineMusicalSymbolOligonArchaion

\u{1d015}: '𝀕'

ByzantineMusicalSymbolGorgonArchaion

\u{1d016}: '𝀖'

ByzantineMusicalSymbolPsilon

\u{1d017}: '𝀗'

ByzantineMusicalSymbolChamilon

\u{1d018}: '𝀘'

ByzantineMusicalSymbolVathy

\u{1d019}: '𝀙'

ByzantineMusicalSymbolIsonArchaion

\u{1d01a}: '𝀚'

ByzantineMusicalSymbolKentimaArchaion

\u{1d01b}: '𝀛'

ByzantineMusicalSymbolKentimataArchaion

\u{1d01c}: '𝀜'

ByzantineMusicalSymbolSaximata

\u{1d01d}: '𝀝'

ByzantineMusicalSymbolParichon

\u{1d01e}: '𝀞'

ByzantineMusicalSymbolStavrosApodexia

\u{1d01f}: '𝀟'

ByzantineMusicalSymbolOxeiaiArchaion

\u{1d020}: '𝀠'

ByzantineMusicalSymbolVareiaiArchaion

\u{1d021}: '𝀡'

ByzantineMusicalSymbolApodermaArchaion

\u{1d022}: '𝀢'

ByzantineMusicalSymbolApothema

\u{1d023}: '𝀣'

ByzantineMusicalSymbolKlasma

\u{1d024}: '𝀤'

ByzantineMusicalSymbolRevma

\u{1d025}: '𝀥'

ByzantineMusicalSymbolPiasmaArchaion

\u{1d026}: '𝀦'

ByzantineMusicalSymbolTinagma

\u{1d027}: '𝀧'

ByzantineMusicalSymbolAnatrichisma

\u{1d028}: '𝀨'

ByzantineMusicalSymbolSeisma

\u{1d029}: '𝀩'

ByzantineMusicalSymbolSynagmaArchaion

\u{1d02a}: '𝀪'

ByzantineMusicalSymbolSynagmaMetaStavrou

\u{1d02b}: '𝀫'

ByzantineMusicalSymbolOyranismaArchaion

\u{1d02c}: '𝀬'

ByzantineMusicalSymbolThema

\u{1d02d}: '𝀭'

ByzantineMusicalSymbolLemoi

\u{1d02e}: '𝀮'

ByzantineMusicalSymbolDyo

\u{1d02f}: '𝀯'

ByzantineMusicalSymbolTria

\u{1d030}: '𝀰'

ByzantineMusicalSymbolTessera

\u{1d031}: '𝀱'

ByzantineMusicalSymbolKratimata

\u{1d032}: '𝀲'

ByzantineMusicalSymbolApesoExoNeo

\u{1d033}: '𝀳'

ByzantineMusicalSymbolFthoraArchaion

\u{1d034}: '𝀴'

ByzantineMusicalSymbolImifthora

\u{1d035}: '𝀵'

ByzantineMusicalSymbolTromikonArchaion

\u{1d036}: '𝀶'

ByzantineMusicalSymbolKatavaTromikon

\u{1d037}: '𝀷'

ByzantineMusicalSymbolPelaston

\u{1d038}: '𝀸'

ByzantineMusicalSymbolPsifiston

\u{1d039}: '𝀹'

ByzantineMusicalSymbolKontevma

\u{1d03a}: '𝀺'

ByzantineMusicalSymbolChorevmaArchaion

\u{1d03b}: '𝀻'

ByzantineMusicalSymbolRapisma

\u{1d03c}: '𝀼'

ByzantineMusicalSymbolParakalesmaArchaion

\u{1d03d}: '𝀽'

ByzantineMusicalSymbolParaklitikiArchaion

\u{1d03e}: '𝀾'

ByzantineMusicalSymbolIchadin

\u{1d03f}: '𝀿'

ByzantineMusicalSymbolNana

\u{1d040}: '𝁀'

ByzantineMusicalSymbolPetasma

\u{1d041}: '𝁁'

ByzantineMusicalSymbolKontevmaAllo

\u{1d042}: '𝁂'

ByzantineMusicalSymbolTromikonAllo

\u{1d043}: '𝁃'

ByzantineMusicalSymbolStraggismata

\u{1d044}: '𝁄'

ByzantineMusicalSymbolGronthismata

\u{1d045}: '𝁅'

ByzantineMusicalSymbolIsonNeo

\u{1d046}: '𝁆'

ByzantineMusicalSymbolOligonNeo

\u{1d047}: '𝁇'

ByzantineMusicalSymbolOxeiaNeo

\u{1d048}: '𝁈'

ByzantineMusicalSymbolPetasti

\u{1d049}: '𝁉'

ByzantineMusicalSymbolKoufisma

\u{1d04a}: '𝁊'

ByzantineMusicalSymbolPetastokoufisma

\u{1d04b}: '𝁋'

ByzantineMusicalSymbolKratimokoufisma

\u{1d04c}: '𝁌'

ByzantineMusicalSymbolPelastonNeo

\u{1d04d}: '𝁍'

ByzantineMusicalSymbolKentimataNeoAno

\u{1d04e}: '𝁎'

ByzantineMusicalSymbolKentimaNeoAno

\u{1d04f}: '𝁏'

ByzantineMusicalSymbolYpsili

\u{1d050}: '𝁐'

ByzantineMusicalSymbolApostrofosNeo

\u{1d051}: '𝁑'

ByzantineMusicalSymbolApostrofoiSyndesmosNeo

\u{1d052}: '𝁒'

ByzantineMusicalSymbolYporroi

\u{1d053}: '𝁓'

ByzantineMusicalSymbolKratimoyporroon

\u{1d054}: '𝁔'

ByzantineMusicalSymbolElafron

\u{1d055}: '𝁕'

ByzantineMusicalSymbolChamili

\u{1d056}: '𝁖'

ByzantineMusicalSymbolMikronIson

\u{1d057}: '𝁗'

ByzantineMusicalSymbolVareiaNeo

\u{1d058}: '𝁘'

ByzantineMusicalSymbolPiasmaNeo

\u{1d059}: '𝁙'

ByzantineMusicalSymbolPsifistonNeo

\u{1d05a}: '𝁚'

ByzantineMusicalSymbolOmalon

\u{1d05b}: '𝁛'

ByzantineMusicalSymbolAntikenoma

\u{1d05c}: '𝁜'

ByzantineMusicalSymbolLygisma

\u{1d05d}: '𝁝'

ByzantineMusicalSymbolParaklitikiNeo

\u{1d05e}: '𝁞'

ByzantineMusicalSymbolParakalesmaNeo

\u{1d05f}: '𝁟'

ByzantineMusicalSymbolEteronParakalesma

\u{1d060}: '𝁠'

ByzantineMusicalSymbolKylisma

\u{1d061}: '𝁡'

ByzantineMusicalSymbolAntikenokylisma

\u{1d062}: '𝁢'

ByzantineMusicalSymbolTromikonNeo

\u{1d063}: '𝁣'

ByzantineMusicalSymbolEkstrepton

\u{1d064}: '𝁤'

ByzantineMusicalSymbolSynagmaNeo

\u{1d065}: '𝁥'

ByzantineMusicalSymbolSyrma

\u{1d066}: '𝁦'

ByzantineMusicalSymbolChorevmaNeo

\u{1d067}: '𝁧'

ByzantineMusicalSymbolEpegerma

\u{1d068}: '𝁨'

ByzantineMusicalSymbolSeismaNeo

\u{1d069}: '𝁩'

ByzantineMusicalSymbolXironKlasma

\u{1d06a}: '𝁪'

ByzantineMusicalSymbolTromikopsifiston

\u{1d06b}: '𝁫'

ByzantineMusicalSymbolPsifistolygisma

\u{1d06c}: '𝁬'

ByzantineMusicalSymbolTromikolygisma

\u{1d06d}: '𝁭'

ByzantineMusicalSymbolTromikoparakalesma

\u{1d06e}: '𝁮'

ByzantineMusicalSymbolPsifistoparakalesma

\u{1d06f}: '𝁯'

ByzantineMusicalSymbolTromikosynagma

\u{1d070}: '𝁰'

ByzantineMusicalSymbolPsifistosynagma

\u{1d071}: '𝁱'

ByzantineMusicalSymbolGorgosyntheton

\u{1d072}: '𝁲'

ByzantineMusicalSymbolArgosyntheton

\u{1d073}: '𝁳'

ByzantineMusicalSymbolEteronArgosyntheton

\u{1d074}: '𝁴'

ByzantineMusicalSymbolOyranismaNeo

\u{1d075}: '𝁵'

ByzantineMusicalSymbolThematismosEso

\u{1d076}: '𝁶'

ByzantineMusicalSymbolThematismosExo

\u{1d077}: '𝁷'

ByzantineMusicalSymbolThemaAploun

\u{1d078}: '𝁸'

ByzantineMusicalSymbolThesKaiApothes

\u{1d079}: '𝁹'

ByzantineMusicalSymbolKatavasma

\u{1d07a}: '𝁺'

ByzantineMusicalSymbolEndofonon

\u{1d07b}: '𝁻'

ByzantineMusicalSymbolYfenKato

\u{1d07c}: '𝁼'

ByzantineMusicalSymbolYfenAno

\u{1d07d}: '𝁽'

ByzantineMusicalSymbolStavros

\u{1d07e}: '𝁾'

ByzantineMusicalSymbolKlasmaAno

\u{1d07f}: '𝁿'

ByzantineMusicalSymbolDipliArchaion

\u{1d080}: '𝂀'

ByzantineMusicalSymbolKratimaArchaion

\u{1d081}: '𝂁'

ByzantineMusicalSymbolKratimaAllo

\u{1d082}: '𝂂'

ByzantineMusicalSymbolKratimaNeo

\u{1d083}: '𝂃'

ByzantineMusicalSymbolApodermaNeo

\u{1d084}: '𝂄'

ByzantineMusicalSymbolApli

\u{1d085}: '𝂅'

ByzantineMusicalSymbolDipli

\u{1d086}: '𝂆'

ByzantineMusicalSymbolTripli

\u{1d087}: '𝂇'

ByzantineMusicalSymbolTetrapli

\u{1d088}: '𝂈'

ByzantineMusicalSymbolKoronis

\u{1d089}: '𝂉'

ByzantineMusicalSymbolLeimmaEnosChronou

\u{1d08a}: '𝂊'

ByzantineMusicalSymbolLeimmaDyoChronon

\u{1d08b}: '𝂋'

ByzantineMusicalSymbolLeimmaTrionChronon

\u{1d08c}: '𝂌'

ByzantineMusicalSymbolLeimmaTessaronChronon

\u{1d08d}: '𝂍'

ByzantineMusicalSymbolLeimmaImiseosChronou

\u{1d08e}: '𝂎'

ByzantineMusicalSymbolGorgonNeoAno

\u{1d08f}: '𝂏'

ByzantineMusicalSymbolGorgonParestigmenonAristera

\u{1d090}: '𝂐'

ByzantineMusicalSymbolGorgonParestigmenonDexia

\u{1d091}: '𝂑'

ByzantineMusicalSymbolDigorgon

\u{1d092}: '𝂒'

ByzantineMusicalSymbolDigorgonParestigmenonAristeraKato

\u{1d093}: '𝂓'

ByzantineMusicalSymbolDigorgonParestigmenonAristeraAno

\u{1d094}: '𝂔'

ByzantineMusicalSymbolDigorgonParestigmenonDexia

\u{1d095}: '𝂕'

ByzantineMusicalSymbolTrigorgon

\u{1d096}: '𝂖'

ByzantineMusicalSymbolArgon

\u{1d097}: '𝂗'

ByzantineMusicalSymbolImidiargon

\u{1d098}: '𝂘'

ByzantineMusicalSymbolDiargon

\u{1d099}: '𝂙'

ByzantineMusicalSymbolAgogiPoliArgi

\u{1d09a}: '𝂚'

ByzantineMusicalSymbolAgogiArgoteri

\u{1d09b}: '𝂛'

ByzantineMusicalSymbolAgogiArgi

\u{1d09c}: '𝂜'

ByzantineMusicalSymbolAgogiMetria

\u{1d09d}: '𝂝'

ByzantineMusicalSymbolAgogiMesi

\u{1d09e}: '𝂞'

ByzantineMusicalSymbolAgogiGorgi

\u{1d09f}: '𝂟'

ByzantineMusicalSymbolAgogiGorgoteri

\u{1d0a0}: '𝂠'

ByzantineMusicalSymbolAgogiPoliGorgi

\u{1d0a1}: '𝂡'

ByzantineMusicalSymbolMartyriaProtosIchos

\u{1d0a2}: '𝂢'

ByzantineMusicalSymbolMartyriaAlliProtosIchos

\u{1d0a3}: '𝂣'

ByzantineMusicalSymbolMartyriaDeyterosIchos

\u{1d0a4}: '𝂤'

ByzantineMusicalSymbolMartyriaAlliDeyterosIchos

\u{1d0a5}: '𝂥'

ByzantineMusicalSymbolMartyriaTritosIchos

\u{1d0a6}: '𝂦'

ByzantineMusicalSymbolMartyriaTrifonias

\u{1d0a7}: '𝂧'

ByzantineMusicalSymbolMartyriaTetartosIchos

\u{1d0a8}: '𝂨'

ByzantineMusicalSymbolMartyriaTetartosLegetosIchos

\u{1d0a9}: '𝂩'

ByzantineMusicalSymbolMartyriaLegetosIchos

\u{1d0aa}: '𝂪'

ByzantineMusicalSymbolMartyriaPlagiosIchos

\u{1d0ab}: '𝂫'

ByzantineMusicalSymbolIsakiaTelousIchimatos

\u{1d0ac}: '𝂬'

ByzantineMusicalSymbolApostrofoiTelousIchimatos

\u{1d0ad}: '𝂭'

ByzantineMusicalSymbolFanerosisTetrafonias

\u{1d0ae}: '𝂮'

ByzantineMusicalSymbolFanerosisMonofonias

\u{1d0af}: '𝂯'

ByzantineMusicalSymbolFanerosisDifonias

\u{1d0b0}: '𝂰'

ByzantineMusicalSymbolMartyriaVarysIchos

\u{1d0b1}: '𝂱'

ByzantineMusicalSymbolMartyriaProtovarysIchos

\u{1d0b2}: '𝂲'

ByzantineMusicalSymbolMartyriaPlagiosTetartosIchos

\u{1d0b3}: '𝂳'

ByzantineMusicalSymbolGorthmikonNAploun

\u{1d0b4}: '𝂴'

ByzantineMusicalSymbolGorthmikonNDiploun

\u{1d0b5}: '𝂵'

ByzantineMusicalSymbolEnarxisKaiFthoraVou

\u{1d0b6}: '𝂶'

ByzantineMusicalSymbolImifonon

\u{1d0b7}: '𝂷'

ByzantineMusicalSymbolImifthoron

\u{1d0b8}: '𝂸'

ByzantineMusicalSymbolFthoraArchaionDeyterouIchou

\u{1d0b9}: '𝂹'

ByzantineMusicalSymbolFthoraDiatonikiPa

\u{1d0ba}: '𝂺'

ByzantineMusicalSymbolFthoraDiatonikiNana

\u{1d0bb}: '𝂻'

ByzantineMusicalSymbolFthoraNaosIchos

\u{1d0bc}: '𝂼'

ByzantineMusicalSymbolFthoraDiatonikiDi

\u{1d0bd}: '𝂽'

ByzantineMusicalSymbolFthoraSklironDiatononDi

\u{1d0be}: '𝂾'

ByzantineMusicalSymbolFthoraDiatonikiKe

\u{1d0bf}: '𝂿'

ByzantineMusicalSymbolFthoraDiatonikiZo

\u{1d0c0}: '𝃀'

ByzantineMusicalSymbolFthoraDiatonikiNiKato

\u{1d0c1}: '𝃁'

ByzantineMusicalSymbolFthoraDiatonikiNiAno

\u{1d0c2}: '𝃂'

ByzantineMusicalSymbolFthoraMalakonChromaDifonias

\u{1d0c3}: '𝃃'

ByzantineMusicalSymbolFthoraMalakonChromaMonofonias

\u{1d0c4}: '𝃄'

ByzantineMusicalSymbolFhtoraSklironChromaVasis

\u{1d0c5}: '𝃅'

ByzantineMusicalSymbolFthoraSklironChromaSynafi

\u{1d0c6}: '𝃆'

ByzantineMusicalSymbolFthoraNenano

\u{1d0c7}: '𝃇'

ByzantineMusicalSymbolChroaZygos

\u{1d0c8}: '𝃈'

ByzantineMusicalSymbolChroaKliton

\u{1d0c9}: '𝃉'

ByzantineMusicalSymbolChroaSpathi

\u{1d0ca}: '𝃊'

ByzantineMusicalSymbolFthoraIYfesisTetartimorion

\u{1d0cb}: '𝃋'

ByzantineMusicalSymbolFthoraEnarmoniosAntifonia

\u{1d0cc}: '𝃌'

ByzantineMusicalSymbolYfesisTritimorion

\u{1d0cd}: '𝃍'

ByzantineMusicalSymbolDiesisTritimorion

\u{1d0ce}: '𝃎'

ByzantineMusicalSymbolDiesisTetartimorion

\u{1d0cf}: '𝃏'

ByzantineMusicalSymbolDiesisApliDyoDodekata

\u{1d0d0}: '𝃐'

ByzantineMusicalSymbolDiesisMonogrammosTesseraDodekata

\u{1d0d1}: '𝃑'

ByzantineMusicalSymbolDiesisDigrammosExDodekata

\u{1d0d2}: '𝃒'

ByzantineMusicalSymbolDiesisTrigrammosOktoDodekata

\u{1d0d3}: '𝃓'

ByzantineMusicalSymbolYfesisApliDyoDodekata

\u{1d0d4}: '𝃔'

ByzantineMusicalSymbolYfesisMonogrammosTesseraDodekata

\u{1d0d5}: '𝃕'

ByzantineMusicalSymbolYfesisDigrammosExDodekata

\u{1d0d6}: '𝃖'

ByzantineMusicalSymbolYfesisTrigrammosOktoDodekata

\u{1d0d7}: '𝃗'

ByzantineMusicalSymbolGenikiDiesis

\u{1d0d8}: '𝃘'

ByzantineMusicalSymbolGenikiYfesis

\u{1d0d9}: '𝃙'

ByzantineMusicalSymbolDiastoliApliMikri

\u{1d0da}: '𝃚'

ByzantineMusicalSymbolDiastoliApliMegali

\u{1d0db}: '𝃛'

ByzantineMusicalSymbolDiastoliDipli

\u{1d0dc}: '𝃜'

ByzantineMusicalSymbolDiastoliTheseos

\u{1d0dd}: '𝃝'

ByzantineMusicalSymbolSimansisTheseos

\u{1d0de}: '𝃞'

ByzantineMusicalSymbolSimansisTheseosDisimou

\u{1d0df}: '𝃟'

ByzantineMusicalSymbolSimansisTheseosTrisimou

\u{1d0e0}: '𝃠'

ByzantineMusicalSymbolSimansisTheseosTetrasimou

\u{1d0e1}: '𝃡'

ByzantineMusicalSymbolSimansisArseos

\u{1d0e2}: '𝃢'

ByzantineMusicalSymbolSimansisArseosDisimou

\u{1d0e3}: '𝃣'

ByzantineMusicalSymbolSimansisArseosTrisimou

\u{1d0e4}: '𝃤'

ByzantineMusicalSymbolSimansisArseosTetrasimou

\u{1d0e5}: '𝃥'

ByzantineMusicalSymbolDigrammaGg

\u{1d0e6}: '𝃦'

ByzantineMusicalSymbolDiftoggosOu

\u{1d0e7}: '𝃧'

ByzantineMusicalSymbolStigma

\u{1d0e8}: '𝃨'

ByzantineMusicalSymbolArktikoPa

\u{1d0e9}: '𝃩'

ByzantineMusicalSymbolArktikoVou

\u{1d0ea}: '𝃪'

ByzantineMusicalSymbolArktikoGa

\u{1d0eb}: '𝃫'

ByzantineMusicalSymbolArktikoDi

\u{1d0ec}: '𝃬'

ByzantineMusicalSymbolArktikoKe

\u{1d0ed}: '𝃭'

ByzantineMusicalSymbolArktikoZo

\u{1d0ee}: '𝃮'

ByzantineMusicalSymbolArktikoNi

\u{1d0ef}: '𝃯'

ByzantineMusicalSymbolKentimataNeoMeso

\u{1d0f0}: '𝃰'

ByzantineMusicalSymbolKentimaNeoMeso

\u{1d0f1}: '𝃱'

ByzantineMusicalSymbolKentimataNeoKato

\u{1d0f2}: '𝃲'

ByzantineMusicalSymbolKentimaNeoKato

\u{1d0f3}: '𝃳'

ByzantineMusicalSymbolKlasmaKato

\u{1d0f4}: '𝃴'

ByzantineMusicalSymbolGorgonNeoKato

\u{1d0f5}: '𝃵'

Methods

impl ByzantineMusicalSymbols

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case