Enum unicode_types::chess_symbols::ChessSymbols

pub enum ChessSymbols {
    NeutralChessKing,
    NeutralChessQueen,
    NeutralChessRook,
    NeutralChessBishop,
    NeutralChessKnight,
    NeutralChessPawn,
    WhiteChessKnightRotatedFortyDashFiveDegrees,
    BlackChessKnightRotatedFortyDashFiveDegrees,
    NeutralChessKnightRotatedFortyDashFiveDegrees,
    WhiteChessKingRotatedNinetyDegrees,
    WhiteChessQueenRotatedNinetyDegrees,
    WhiteChessRookRotatedNinetyDegrees,
    WhiteChessBishopRotatedNinetyDegrees,
    WhiteChessKnightRotatedNinetyDegrees,
    WhiteChessPawnRotatedNinetyDegrees,
    BlackChessKingRotatedNinetyDegrees,
    BlackChessQueenRotatedNinetyDegrees,
    BlackChessRookRotatedNinetyDegrees,
    BlackChessBishopRotatedNinetyDegrees,
    BlackChessKnightRotatedNinetyDegrees,
    BlackChessPawnRotatedNinetyDegrees,
    NeutralChessKingRotatedNinetyDegrees,
    NeutralChessQueenRotatedNinetyDegrees,
    NeutralChessRookRotatedNinetyDegrees,
    NeutralChessBishopRotatedNinetyDegrees,
    NeutralChessKnightRotatedNinetyDegrees,
    NeutralChessPawnRotatedNinetyDegrees,
    WhiteChessKnightRotatedOneHundredThirtyDashFiveDegrees,
    BlackChessKnightRotatedOneHundredThirtyDashFiveDegrees,
    NeutralChessKnightRotatedOneHundredThirtyDashFiveDegrees,
    WhiteChessTurnedKing,
    WhiteChessTurnedQueen,
    WhiteChessTurnedRook,
    WhiteChessTurnedBishop,
    WhiteChessTurnedKnight,
    WhiteChessTurnedPawn,
    BlackChessTurnedKing,
    BlackChessTurnedQueen,
    BlackChessTurnedRook,
    BlackChessTurnedBishop,
    BlackChessTurnedKnight,
    BlackChessTurnedPawn,
    NeutralChessTurnedKing,
    NeutralChessTurnedQueen,
    NeutralChessTurnedRook,
    NeutralChessTurnedBishop,
    NeutralChessTurnedKnight,
    NeutralChessTurnedPawn,
    WhiteChessKnightRotatedTwoHundredTwentyDashFiveDegrees,
    BlackChessKnightRotatedTwoHundredTwentyDashFiveDegrees,
    NeutralChessKnightRotatedTwoHundredTwentyDashFiveDegrees,
    WhiteChessKingRotatedTwoHundredSeventyDegrees,
    WhiteChessQueenRotatedTwoHundredSeventyDegrees,
    WhiteChessRookRotatedTwoHundredSeventyDegrees,
    WhiteChessBishopRotatedTwoHundredSeventyDegrees,
    WhiteChessKnightRotatedTwoHundredSeventyDegrees,
    WhiteChessPawnRotatedTwoHundredSeventyDegrees,
    BlackChessKingRotatedTwoHundredSeventyDegrees,
    BlackChessQueenRotatedTwoHundredSeventyDegrees,
    BlackChessRookRotatedTwoHundredSeventyDegrees,
    BlackChessBishopRotatedTwoHundredSeventyDegrees,
    BlackChessKnightRotatedTwoHundredSeventyDegrees,
    BlackChessPawnRotatedTwoHundredSeventyDegrees,
    NeutralChessKingRotatedTwoHundredSeventyDegrees,
    NeutralChessQueenRotatedTwoHundredSeventyDegrees,
    NeutralChessRookRotatedTwoHundredSeventyDegrees,
    NeutralChessBishopRotatedTwoHundredSeventyDegrees,
    NeutralChessKnightRotatedTwoHundredSeventyDegrees,
    NeutralChessPawnRotatedTwoHundredSeventyDegrees,
    WhiteChessKnightRotatedThreeHundredFifteenDegrees,
    BlackChessKnightRotatedThreeHundredFifteenDegrees,
    NeutralChessKnightRotatedThreeHundredFifteenDegrees,
    WhiteChessEquihopper,
    BlackChessEquihopper,
    NeutralChessEquihopper,
    WhiteChessEquihopperRotatedNinetyDegrees,
    BlackChessEquihopperRotatedNinetyDegrees,
    NeutralChessEquihopperRotatedNinetyDegrees,
    WhiteChessKnightDashQueen,
    WhiteChessKnightDashRook,
    WhiteChessKnightDashBishop,
    BlackChessKnightDashQueen,
    BlackChessKnightDashRook,
    BlackChessKnightDashBishop,
    XiangqiRedGeneral,
    XiangqiRedMandarin,
    XiangqiRedElephant,
    XiangqiRedHorse,
    XiangqiRedChariot,
    XiangqiRedCannon,
    XiangqiRedSoldier,
    XiangqiBlackGeneral,
    XiangqiBlackMandarin,
    XiangqiBlackElephant,
    XiangqiBlackHorse,
    XiangqiBlackChariot,
    XiangqiBlackCannon,
    XiangqiBlackSoldier,
}

An enum to represent all characters in the ChessSymbols block.

Variants
NeutralChessKing

\u{1fa00}: 'ğŸ¨€'

NeutralChessQueen

\u{1fa01}: 'ğŸ¨'

NeutralChessRook

\u{1fa02}: 'ğŸ¨‚'

NeutralChessBishop

\u{1fa03}: 'ğŸ¨ƒ'

NeutralChessKnight

\u{1fa04}: 'ğŸ¨„'

NeutralChessPawn

\u{1fa05}: 'ğŸ¨…'

WhiteChessKnightRotatedFortyDashFiveDegrees

\u{1fa06}: 'ğŸ¨†'

BlackChessKnightRotatedFortyDashFiveDegrees

\u{1fa07}: 'ğŸ¨‡'

NeutralChessKnightRotatedFortyDashFiveDegrees

\u{1fa08}: 'ğŸ¨ˆ'

WhiteChessKingRotatedNinetyDegrees

\u{1fa09}: 'ğŸ¨‰'

WhiteChessQueenRotatedNinetyDegrees

\u{1fa0a}: 'ğŸ¨Š'

WhiteChessRookRotatedNinetyDegrees

\u{1fa0b}: 'ğŸ¨‹'

WhiteChessBishopRotatedNinetyDegrees

\u{1fa0c}: 'ğŸ¨Œ'

WhiteChessKnightRotatedNinetyDegrees

\u{1fa0d}: 'ğŸ¨'

WhiteChessPawnRotatedNinetyDegrees

\u{1fa0e}: 'ğŸ¨'

BlackChessKingRotatedNinetyDegrees

\u{1fa0f}: 'ğŸ¨'

BlackChessQueenRotatedNinetyDegrees

\u{1fa10}: 'ğŸ¨'

BlackChessRookRotatedNinetyDegrees

\u{1fa11}: 'ğŸ¨‘'

BlackChessBishopRotatedNinetyDegrees

\u{1fa12}: 'ğŸ¨’'

BlackChessKnightRotatedNinetyDegrees

\u{1fa13}: 'ğŸ¨“'

BlackChessPawnRotatedNinetyDegrees

\u{1fa14}: 'ğŸ¨”'

NeutralChessKingRotatedNinetyDegrees

\u{1fa15}: 'ğŸ¨•'

NeutralChessQueenRotatedNinetyDegrees

\u{1fa16}: 'ğŸ¨–'

NeutralChessRookRotatedNinetyDegrees

\u{1fa17}: 'ğŸ¨—'

NeutralChessBishopRotatedNinetyDegrees

\u{1fa18}: 'ğŸ¨˜'

NeutralChessKnightRotatedNinetyDegrees

\u{1fa19}: 'ğŸ¨™'

NeutralChessPawnRotatedNinetyDegrees

\u{1fa1a}: 'ğŸ¨š'

WhiteChessKnightRotatedOneHundredThirtyDashFiveDegrees

\u{1fa1b}: 'ğŸ¨›'

BlackChessKnightRotatedOneHundredThirtyDashFiveDegrees

\u{1fa1c}: 'ğŸ¨œ'

NeutralChessKnightRotatedOneHundredThirtyDashFiveDegrees

\u{1fa1d}: 'ğŸ¨'

WhiteChessTurnedKing

\u{1fa1e}: 'ğŸ¨'

WhiteChessTurnedQueen

\u{1fa1f}: 'ğŸ¨Ÿ'

WhiteChessTurnedRook

\u{1fa20}: 'ğŸ¨ '

WhiteChessTurnedBishop

\u{1fa21}: 'ğŸ¨¡'

WhiteChessTurnedKnight

\u{1fa22}: 'ğŸ¨¢'

WhiteChessTurnedPawn

\u{1fa23}: 'ğŸ¨£'

BlackChessTurnedKing

\u{1fa24}: 'ğŸ¨¤'

BlackChessTurnedQueen

\u{1fa25}: 'ğŸ¨¥'

BlackChessTurnedRook

\u{1fa26}: 'ğŸ¨¦'

BlackChessTurnedBishop

\u{1fa27}: 'ğŸ¨'

BlackChessTurnedKnight

\u{1fa28}: 'ğŸ¨¨'

BlackChessTurnedPawn

\u{1fa29}: 'ğŸ¨©'

NeutralChessTurnedKing

\u{1fa2a}: 'ğŸ¨ª'

NeutralChessTurnedQueen

\u{1fa2b}: 'ğŸ¨«'

NeutralChessTurnedRook

\u{1fa2c}: 'ğŸ¨¬'

NeutralChessTurnedBishop

\u{1fa2d}: 'ğŸ¨­'

NeutralChessTurnedKnight

\u{1fa2e}: 'ğŸ¨®'

NeutralChessTurnedPawn

\u{1fa2f}: 'ğŸ¨¯'

WhiteChessKnightRotatedTwoHundredTwentyDashFiveDegrees

\u{1fa30}: 'ğŸ¨°'

BlackChessKnightRotatedTwoHundredTwentyDashFiveDegrees

\u{1fa31}: 'ğŸ¨±'

NeutralChessKnightRotatedTwoHundredTwentyDashFiveDegrees

\u{1fa32}: 'ğŸ¨²'

WhiteChessKingRotatedTwoHundredSeventyDegrees

\u{1fa33}: 'ğŸ¨³'

WhiteChessQueenRotatedTwoHundredSeventyDegrees

\u{1fa34}: 'ğŸ¨´'

WhiteChessRookRotatedTwoHundredSeventyDegrees

\u{1fa35}: 'ğŸ¨µ'

WhiteChessBishopRotatedTwoHundredSeventyDegrees

\u{1fa36}: 'ğŸ¨¶'

WhiteChessKnightRotatedTwoHundredSeventyDegrees

\u{1fa37}: 'ğŸ¨·'

WhiteChessPawnRotatedTwoHundredSeventyDegrees

\u{1fa38}: 'ğŸ¨¸'

BlackChessKingRotatedTwoHundredSeventyDegrees

\u{1fa39}: 'ğŸ¨¹'

BlackChessQueenRotatedTwoHundredSeventyDegrees

\u{1fa3a}: 'ğŸ¨º'

BlackChessRookRotatedTwoHundredSeventyDegrees

\u{1fa3b}: 'ğŸ¨»'

BlackChessBishopRotatedTwoHundredSeventyDegrees

\u{1fa3c}: 'ğŸ¨¼'

BlackChessKnightRotatedTwoHundredSeventyDegrees

\u{1fa3d}: 'ğŸ¨½'

BlackChessPawnRotatedTwoHundredSeventyDegrees

\u{1fa3e}: 'ğŸ¨¾'

NeutralChessKingRotatedTwoHundredSeventyDegrees

\u{1fa3f}: 'ğŸ¨¿'

NeutralChessQueenRotatedTwoHundredSeventyDegrees

\u{1fa40}: 'ğŸ©€'

NeutralChessRookRotatedTwoHundredSeventyDegrees

\u{1fa41}: 'ğŸ©'

NeutralChessBishopRotatedTwoHundredSeventyDegrees

\u{1fa42}: 'ğŸ©‚'

NeutralChessKnightRotatedTwoHundredSeventyDegrees

\u{1fa43}: 'ğŸ©ƒ'

NeutralChessPawnRotatedTwoHundredSeventyDegrees

\u{1fa44}: 'ğŸ©„'

WhiteChessKnightRotatedThreeHundredFifteenDegrees

\u{1fa45}: 'ğŸ©…'

BlackChessKnightRotatedThreeHundredFifteenDegrees

\u{1fa46}: 'ğŸ©†'

NeutralChessKnightRotatedThreeHundredFifteenDegrees

\u{1fa47}: 'ğŸ©‡'

WhiteChessEquihopper

\u{1fa48}: 'ğŸ©ˆ'

BlackChessEquihopper

\u{1fa49}: 'ğŸ©‰'

NeutralChessEquihopper

\u{1fa4a}: 'ğŸ©Š'

WhiteChessEquihopperRotatedNinetyDegrees

\u{1fa4b}: 'ğŸ©‹'

BlackChessEquihopperRotatedNinetyDegrees

\u{1fa4c}: 'ğŸ©Œ'

NeutralChessEquihopperRotatedNinetyDegrees

\u{1fa4d}: 'ğŸ©'

WhiteChessKnightDashQueen

\u{1fa4e}: 'ğŸ©'

WhiteChessKnightDashRook

\u{1fa4f}: 'ğŸ©'

WhiteChessKnightDashBishop

\u{1fa50}: 'ğŸ©'

BlackChessKnightDashQueen

\u{1fa51}: 'ğŸ©‘'

BlackChessKnightDashRook

\u{1fa52}: 'ğŸ©’'

BlackChessKnightDashBishop

\u{1fa53}: 'ğŸ©“'

XiangqiRedGeneral

\u{1fa60}: 'ğŸ© '

XiangqiRedMandarin

\u{1fa61}: 'ğŸ©¡'

XiangqiRedElephant

\u{1fa62}: 'ğŸ©¢'

XiangqiRedHorse

\u{1fa63}: 'ğŸ©£'

XiangqiRedChariot

\u{1fa64}: 'ğŸ©¤'

XiangqiRedCannon

\u{1fa65}: 'ğŸ©¥'

XiangqiRedSoldier

\u{1fa66}: 'ğŸ©¦'

XiangqiBlackGeneral

\u{1fa67}: 'ğŸ©'

XiangqiBlackMandarin

\u{1fa68}: 'ğŸ©¨'

XiangqiBlackElephant

\u{1fa69}: 'ğŸ©©'

XiangqiBlackHorse

\u{1fa6a}: 'ğŸ©ª'

XiangqiBlackChariot

\u{1fa6b}: 'ğŸ©«'

XiangqiBlackCannon

\u{1fa6c}: 'ğŸ©¬'

XiangqiBlackSoldier

\u{1fa6d}: 'ğŸ©­'

Methods

impl ChessSymbols

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case