Enum unicode_types::cjk_compatibility::CJKCompatibility

pub enum CJKCompatibility {
    SquareApaato,
    SquareAruhua,
    SquareAnpea,
    SquareAaru,
    SquareIningu,
    SquareInti,
    SquareUon,
    SquareEsukuudo,
    SquareEekaa,
    SquareOnsu,
    SquareOomu,
    SquareKairi,
    SquareKaratto,
    SquareKarorii,
    SquareGaron,
    SquareGanma,
    SquareGiga,
    SquareGinii,
    SquareKyurii,
    SquareGirudaa,
    SquareKiro,
    SquareKiroguramu,
    SquareKiromeetoru,
    SquareKirowatto,
    SquareGuramu,
    SquareGuramuton,
    SquareKuruzeiro,
    SquareKuroone,
    SquareKeesu,
    SquareKoruna,
    SquareKoopo,
    SquareSaikuru,
    SquareSantiimu,
    SquareSiringu,
    SquareSenti,
    SquareSento,
    SquareDaasu,
    SquareDesi,
    SquareDoru,
    SquareTon,
    SquareNano,
    SquareNotto,
    SquareHaitu,
    SquarePaasento,
    SquarePaatu,
    SquareBaareru,
    SquarePiasutoru,
    SquarePikuru,
    SquarePiko,
    SquareBiru,
    SquareHuaraddo,
    SquareHuiito,
    SquareBussyeru,
    SquareHuran,
    SquareHekutaaru,
    SquarePeso,
    SquarePenihi,
    SquareHerutu,
    SquarePensu,
    SquarePeezi,
    SquareBeeta,
    SquarePointo,
    SquareBoruto,
    SquareHon,
    SquarePondo,
    SquareHooru,
    SquareHoon,
    SquareMaikuro,
    SquareMairu,
    SquareMahha,
    SquareMaruku,
    SquareMansyon,
    SquareMikuron,
    SquareMiri,
    SquareMiribaaru,
    SquareMega,
    SquareMegaton,
    SquareMeetoru,
    SquareYaado,
    SquareYaaru,
    SquareYuan,
    SquareRittoru,
    SquareRira,
    SquareRupii,
    SquareRuuburu,
    SquareRemu,
    SquareRentogen,
    SquareWatto,
    IdeographicTelegraphSymbolForHourZero,
    IdeographicTelegraphSymbolForHourOne,
    IdeographicTelegraphSymbolForHourTwo,
    IdeographicTelegraphSymbolForHourThree,
    IdeographicTelegraphSymbolForHourFour,
    IdeographicTelegraphSymbolForHourFive,
    IdeographicTelegraphSymbolForHourSix,
    IdeographicTelegraphSymbolForHourSeven,
    IdeographicTelegraphSymbolForHourEight,
    IdeographicTelegraphSymbolForHourNine,
    IdeographicTelegraphSymbolForHourTen,
    IdeographicTelegraphSymbolForHourEleven,
    IdeographicTelegraphSymbolForHourTwelve,
    IdeographicTelegraphSymbolForHourThirteen,
    IdeographicTelegraphSymbolForHourFourteen,
    IdeographicTelegraphSymbolForHourFifteen,
    IdeographicTelegraphSymbolForHourSixteen,
    IdeographicTelegraphSymbolForHourSeventeen,
    IdeographicTelegraphSymbolForHourEighteen,
    IdeographicTelegraphSymbolForHourNineteen,
    IdeographicTelegraphSymbolForHourTwenty,
    IdeographicTelegraphSymbolForHourTwentyDashOne,
    IdeographicTelegraphSymbolForHourTwentyDashTwo,
    IdeographicTelegraphSymbolForHourTwentyDashThree,
    IdeographicTelegraphSymbolForHourTwentyDashFour,
    SquareHpa,
    SquareDa,
    SquareAu,
    SquareBar,
    SquareOv,
    SquarePc,
    SquareDm,
    SquareDmSquared,
    SquareDmCubed,
    SquareIu,
    SquareEraNameHeisei,
    SquareEraNameSyouwa,
    SquareEraNameTaisyou,
    SquareEraNameMeizi,
    SquareCorporation,
    SquarePaAmps,
    SquareNa,
    SquareMuA,
    SquareMa,
    SquareKa,
    SquareKb,
    SquareMb,
    SquareGb,
    SquareCal,
    SquareKcal,
    SquarePf,
    SquareNf,
    SquareMuF,
    SquareMuG,
    SquareMg,
    SquareKg,
    SquareHz,
    SquareKhz,
    SquareMhz,
    SquareGhz,
    SquareThz,
    SquareMuL,
    SquareMl,
    SquareDl,
    SquareKl,
    SquareFm,
    SquareNm,
    SquareMuM,
    SquareMm,
    SquareCm,
    SquareKm,
    SquareMmSquared,
    SquareCmSquared,
    SquareMSquared,
    SquareKmSquared,
    SquareMmCubed,
    SquareCmCubed,
    SquareMCubed,
    SquareKmCubed,
    SquareMOverS,
    SquareMOverSSquared,
    SquarePa,
    SquareKpa,
    SquareMpa,
    SquareGpa,
    SquareRad,
    SquareRadOverS,
    SquareRadOverSSquared,
    SquarePs,
    SquareNs,
    SquareMuS,
    SquareMs,
    SquarePv,
    SquareNv,
    SquareMuV,
    SquareMv,
    SquareKv,
    SquareMvMega,
    SquarePw,
    SquareNw,
    SquareMuW,
    SquareMw,
    SquareKw,
    SquareMwMega,
    SquareKOhm,
    SquareMOhm,
    SquareAm,
    SquareBq,
    SquareCc,
    SquareCd,
    SquareCOverKg,
    SquareCo,
    SquareDb,
    SquareGy,
    SquareHa,
    SquareHp,
    SquareIn,
    SquareKk,
    SquareKmCapital,
    SquareKt,
    SquareLm,
    SquareLn,
    SquareLog,
    SquareLx,
    SquareMbSmall,
    SquareMil,
    SquareMol,
    SquarePh,
    SquarePm,
    SquarePpm,
    SquarePr,
    SquareSr,
    SquareSv,
    SquareWb,
    SquareVOverM,
    SquareAOverM,
    IdeographicTelegraphSymbolForDayOne,
    IdeographicTelegraphSymbolForDayTwo,
    IdeographicTelegraphSymbolForDayThree,
    IdeographicTelegraphSymbolForDayFour,
    IdeographicTelegraphSymbolForDayFive,
    IdeographicTelegraphSymbolForDaySix,
    IdeographicTelegraphSymbolForDaySeven,
    IdeographicTelegraphSymbolForDayEight,
    IdeographicTelegraphSymbolForDayNine,
    IdeographicTelegraphSymbolForDayTen,
    IdeographicTelegraphSymbolForDayEleven,
    IdeographicTelegraphSymbolForDayTwelve,
    IdeographicTelegraphSymbolForDayThirteen,
    IdeographicTelegraphSymbolForDayFourteen,
    IdeographicTelegraphSymbolForDayFifteen,
    IdeographicTelegraphSymbolForDaySixteen,
    IdeographicTelegraphSymbolForDaySeventeen,
    IdeographicTelegraphSymbolForDayEighteen,
    IdeographicTelegraphSymbolForDayNineteen,
    IdeographicTelegraphSymbolForDayTwenty,
    IdeographicTelegraphSymbolForDayTwentyDashOne,
    IdeographicTelegraphSymbolForDayTwentyDashTwo,
    IdeographicTelegraphSymbolForDayTwentyDashThree,
    IdeographicTelegraphSymbolForDayTwentyDashFour,
    IdeographicTelegraphSymbolForDayTwentyDashFive,
    IdeographicTelegraphSymbolForDayTwentyDashSix,
    IdeographicTelegraphSymbolForDayTwentyDashSeven,
    IdeographicTelegraphSymbolForDayTwentyDashEight,
    IdeographicTelegraphSymbolForDayTwentyDashNine,
    IdeographicTelegraphSymbolForDayThirty,
    IdeographicTelegraphSymbolForDayThirtyDashOne,
}

An enum to represent all characters in the CJKCompatibility block.

Variants
SquareApaato

\u{3300}: '㌀'

SquareAruhua

\u{3301}: '㌁'

SquareAnpea

\u{3302}: '㌂'

SquareAaru

\u{3303}: '㌃'

SquareIningu

\u{3304}: '㌄'

SquareInti

\u{3305}: '㌅'

SquareUon

\u{3306}: '㌆'

SquareEsukuudo

\u{3307}: '㌇'

SquareEekaa

\u{3308}: '㌈'

SquareOnsu

\u{3309}: '㌉'

SquareOomu

\u{330a}: '㌊'

SquareKairi

\u{330b}: '㌋'

SquareKaratto

\u{330c}: '㌌'

SquareKarorii

\u{330d}: '㌍'

SquareGaron

\u{330e}: '㌎'

SquareGanma

\u{330f}: '㌏'

SquareGiga

\u{3310}: '㌐'

SquareGinii

\u{3311}: '㌑'

SquareKyurii

\u{3312}: '㌒'

SquareGirudaa

\u{3313}: '㌓'

SquareKiro

\u{3314}: '㌔'

SquareKiroguramu

\u{3315}: '㌕'

SquareKiromeetoru

\u{3316}: '㌖'

SquareKirowatto

\u{3317}: '㌗'

SquareGuramu

\u{3318}: '㌘'

SquareGuramuton

\u{3319}: '㌙'

SquareKuruzeiro

\u{331a}: '㌚'

SquareKuroone

\u{331b}: '㌛'

SquareKeesu

\u{331c}: '㌜'

SquareKoruna

\u{331d}: '㌝'

SquareKoopo

\u{331e}: '㌞'

SquareSaikuru

\u{331f}: '㌟'

SquareSantiimu

\u{3320}: '㌠'

SquareSiringu

\u{3321}: '㌡'

SquareSenti

\u{3322}: '㌢'

SquareSento

\u{3323}: '㌣'

SquareDaasu

\u{3324}: '㌤'

SquareDesi

\u{3325}: '㌥'

SquareDoru

\u{3326}: '㌦'

SquareTon

\u{3327}: '㌧'

SquareNano

\u{3328}: '㌨'

SquareNotto

\u{3329}: '㌩'

SquareHaitu

\u{332a}: '㌪'

SquarePaasento

\u{332b}: '㌫'

SquarePaatu

\u{332c}: '㌬'

SquareBaareru

\u{332d}: '㌭'

SquarePiasutoru

\u{332e}: '㌮'

SquarePikuru

\u{332f}: '㌯'

SquarePiko

\u{3330}: '㌰'

SquareBiru

\u{3331}: '㌱'

SquareHuaraddo

\u{3332}: '㌲'

SquareHuiito

\u{3333}: '㌳'

SquareBussyeru

\u{3334}: '㌴'

SquareHuran

\u{3335}: '㌵'

SquareHekutaaru

\u{3336}: '㌶'

SquarePeso

\u{3337}: '㌷'

SquarePenihi

\u{3338}: '㌸'

SquareHerutu

\u{3339}: '㌹'

SquarePensu

\u{333a}: '㌺'

SquarePeezi

\u{333b}: '㌻'

SquareBeeta

\u{333c}: '㌼'

SquarePointo

\u{333d}: '㌽'

SquareBoruto

\u{333e}: '㌾'

SquareHon

\u{333f}: '㌿'

SquarePondo

\u{3340}: '㍀'

SquareHooru

\u{3341}: '㍁'

SquareHoon

\u{3342}: '㍂'

SquareMaikuro

\u{3343}: '㍃'

SquareMairu

\u{3344}: '㍄'

SquareMahha

\u{3345}: '㍅'

SquareMaruku

\u{3346}: '㍆'

SquareMansyon

\u{3347}: '㍇'

SquareMikuron

\u{3348}: '㍈'

SquareMiri

\u{3349}: '㍉'

SquareMiribaaru

\u{334a}: '㍊'

SquareMega

\u{334b}: '㍋'

SquareMegaton

\u{334c}: '㍌'

SquareMeetoru

\u{334d}: '㍍'

SquareYaado

\u{334e}: '㍎'

SquareYaaru

\u{334f}: '㍏'

SquareYuan

\u{3350}: '㍐'

SquareRittoru

\u{3351}: '㍑'

SquareRira

\u{3352}: '㍒'

SquareRupii

\u{3353}: '㍓'

SquareRuuburu

\u{3354}: '㍔'

SquareRemu

\u{3355}: '㍕'

SquareRentogen

\u{3356}: '㍖'

SquareWatto

\u{3357}: '㍗'

IdeographicTelegraphSymbolForHourZero

\u{3358}: '㍘'

IdeographicTelegraphSymbolForHourOne

\u{3359}: '㍙'

IdeographicTelegraphSymbolForHourTwo

\u{335a}: '㍚'

IdeographicTelegraphSymbolForHourThree

\u{335b}: '㍛'

IdeographicTelegraphSymbolForHourFour

\u{335c}: '㍜'

IdeographicTelegraphSymbolForHourFive

\u{335d}: '㍝'

IdeographicTelegraphSymbolForHourSix

\u{335e}: '㍞'

IdeographicTelegraphSymbolForHourSeven

\u{335f}: '㍟'

IdeographicTelegraphSymbolForHourEight

\u{3360}: '㍠'

IdeographicTelegraphSymbolForHourNine

\u{3361}: '㍡'

IdeographicTelegraphSymbolForHourTen

\u{3362}: '㍢'

IdeographicTelegraphSymbolForHourEleven

\u{3363}: '㍣'

IdeographicTelegraphSymbolForHourTwelve

\u{3364}: '㍤'

IdeographicTelegraphSymbolForHourThirteen

\u{3365}: '㍥'

IdeographicTelegraphSymbolForHourFourteen

\u{3366}: '㍦'

IdeographicTelegraphSymbolForHourFifteen

\u{3367}: '㍧'

IdeographicTelegraphSymbolForHourSixteen

\u{3368}: '㍨'

IdeographicTelegraphSymbolForHourSeventeen

\u{3369}: '㍩'

IdeographicTelegraphSymbolForHourEighteen

\u{336a}: '㍪'

IdeographicTelegraphSymbolForHourNineteen

\u{336b}: '㍫'

IdeographicTelegraphSymbolForHourTwenty

\u{336c}: '㍬'

IdeographicTelegraphSymbolForHourTwentyDashOne

\u{336d}: '㍭'

IdeographicTelegraphSymbolForHourTwentyDashTwo

\u{336e}: '㍮'

IdeographicTelegraphSymbolForHourTwentyDashThree

\u{336f}: '㍯'

IdeographicTelegraphSymbolForHourTwentyDashFour

\u{3370}: '㍰'

SquareHpa

\u{3371}: '㍱'

SquareDa

\u{3372}: '㍲'

SquareAu

\u{3373}: '㍳'

SquareBar

\u{3374}: '㍴'

SquareOv

\u{3375}: '㍵'

SquarePc

\u{3376}: '㍶'

SquareDm

\u{3377}: '㍷'

SquareDmSquared

\u{3378}: '㍸'

SquareDmCubed

\u{3379}: '㍹'

SquareIu

\u{337a}: '㍺'

SquareEraNameHeisei

\u{337b}: '㍻'

SquareEraNameSyouwa

\u{337c}: '㍼'

SquareEraNameTaisyou

\u{337d}: '㍽'

SquareEraNameMeizi

\u{337e}: '㍾'

SquareCorporation

\u{337f}: '㍿'

SquarePaAmps

\u{3380}: '㎀'

SquareNa

\u{3381}: '㎁'

SquareMuA

\u{3382}: '㎂'

SquareMa

\u{3383}: '㎃'

SquareKa

\u{3384}: '㎄'

SquareKb

\u{3385}: '㎅'

SquareMb

\u{3386}: '㎆'

SquareGb

\u{3387}: '㎇'

SquareCal

\u{3388}: '㎈'

SquareKcal

\u{3389}: '㎉'

SquarePf

\u{338a}: '㎊'

SquareNf

\u{338b}: '㎋'

SquareMuF

\u{338c}: '㎌'

SquareMuG

\u{338d}: '㎍'

SquareMg

\u{338e}: '㎎'

SquareKg

\u{338f}: '㎏'

SquareHz

\u{3390}: '㎐'

SquareKhz

\u{3391}: '㎑'

SquareMhz

\u{3392}: '㎒'

SquareGhz

\u{3393}: '㎓'

SquareThz

\u{3394}: '㎔'

SquareMuL

\u{3395}: '㎕'

SquareMl

\u{3396}: '㎖'

SquareDl

\u{3397}: '㎗'

SquareKl

\u{3398}: '㎘'

SquareFm

\u{3399}: '㎙'

SquareNm

\u{339a}: '㎚'

SquareMuM

\u{339b}: '㎛'

SquareMm

\u{339c}: '㎜'

SquareCm

\u{339d}: '㎝'

SquareKm

\u{339e}: '㎞'

SquareMmSquared

\u{339f}: '㎟'

SquareCmSquared

\u{33a0}: '㎠'

SquareMSquared

\u{33a1}: '㎡'

SquareKmSquared

\u{33a2}: '㎢'

SquareMmCubed

\u{33a3}: '㎣'

SquareCmCubed

\u{33a4}: '㎤'

SquareMCubed

\u{33a5}: '㎥'

SquareKmCubed

\u{33a6}: '㎦'

SquareMOverS

\u{33a7}: '㎧'

SquareMOverSSquared

\u{33a8}: '㎨'

SquarePa

\u{33a9}: '㎩'

SquareKpa

\u{33aa}: '㎪'

SquareMpa

\u{33ab}: '㎫'

SquareGpa

\u{33ac}: '㎬'

SquareRad

\u{33ad}: '㎭'

SquareRadOverS

\u{33ae}: '㎮'

SquareRadOverSSquared

\u{33af}: '㎯'

SquarePs

\u{33b0}: '㎰'

SquareNs

\u{33b1}: '㎱'

SquareMuS

\u{33b2}: '㎲'

SquareMs

\u{33b3}: '㎳'

SquarePv

\u{33b4}: '㎴'

SquareNv

\u{33b5}: '㎵'

SquareMuV

\u{33b6}: '㎶'

SquareMv

\u{33b7}: '㎷'

SquareKv

\u{33b8}: '㎸'

SquareMvMega

\u{33b9}: '㎹'

SquarePw

\u{33ba}: '㎺'

SquareNw

\u{33bb}: '㎻'

SquareMuW

\u{33bc}: '㎼'

SquareMw

\u{33bd}: '㎽'

SquareKw

\u{33be}: '㎾'

SquareMwMega

\u{33bf}: '㎿'

SquareKOhm

\u{33c0}: '㏀'

SquareMOhm

\u{33c1}: '㏁'

SquareAm

\u{33c2}: '㏂'

SquareBq

\u{33c3}: '㏃'

SquareCc

\u{33c4}: '㏄'

SquareCd

\u{33c5}: '㏅'

SquareCOverKg

\u{33c6}: '㏆'

SquareCo

\u{33c7}: '㏇'

SquareDb

\u{33c8}: '㏈'

SquareGy

\u{33c9}: '㏉'

SquareHa

\u{33ca}: '㏊'

SquareHp

\u{33cb}: '㏋'

SquareIn

\u{33cc}: '㏌'

SquareKk

\u{33cd}: '㏍'

SquareKmCapital

\u{33ce}: '㏎'

SquareKt

\u{33cf}: '㏏'

SquareLm

\u{33d0}: '㏐'

SquareLn

\u{33d1}: '㏑'

SquareLog

\u{33d2}: '㏒'

SquareLx

\u{33d3}: '㏓'

SquareMbSmall

\u{33d4}: '㏔'

SquareMil

\u{33d5}: '㏕'

SquareMol

\u{33d6}: '㏖'

SquarePh

\u{33d7}: '㏗'

SquarePm

\u{33d8}: '㏘'

SquarePpm

\u{33d9}: '㏙'

SquarePr

\u{33da}: '㏚'

SquareSr

\u{33db}: '㏛'

SquareSv

\u{33dc}: '㏜'

SquareWb

\u{33dd}: '㏝'

SquareVOverM

\u{33de}: '㏞'

SquareAOverM

\u{33df}: '㏟'

IdeographicTelegraphSymbolForDayOne

\u{33e0}: '㏠'

IdeographicTelegraphSymbolForDayTwo

\u{33e1}: '㏡'

IdeographicTelegraphSymbolForDayThree

\u{33e2}: '㏢'

IdeographicTelegraphSymbolForDayFour

\u{33e3}: '㏣'

IdeographicTelegraphSymbolForDayFive

\u{33e4}: '㏤'

IdeographicTelegraphSymbolForDaySix

\u{33e5}: '㏥'

IdeographicTelegraphSymbolForDaySeven

\u{33e6}: '㏦'

IdeographicTelegraphSymbolForDayEight

\u{33e7}: '㏧'

IdeographicTelegraphSymbolForDayNine

\u{33e8}: '㏨'

IdeographicTelegraphSymbolForDayTen

\u{33e9}: '㏩'

IdeographicTelegraphSymbolForDayEleven

\u{33ea}: '㏪'

IdeographicTelegraphSymbolForDayTwelve

\u{33eb}: '㏫'

IdeographicTelegraphSymbolForDayThirteen

\u{33ec}: '㏬'

IdeographicTelegraphSymbolForDayFourteen

\u{33ed}: '㏭'

IdeographicTelegraphSymbolForDayFifteen

\u{33ee}: '㏮'

IdeographicTelegraphSymbolForDaySixteen

\u{33ef}: '㏯'

IdeographicTelegraphSymbolForDaySeventeen

\u{33f0}: '㏰'

IdeographicTelegraphSymbolForDayEighteen

\u{33f1}: '㏱'

IdeographicTelegraphSymbolForDayNineteen

\u{33f2}: '㏲'

IdeographicTelegraphSymbolForDayTwenty

\u{33f3}: '㏳'

IdeographicTelegraphSymbolForDayTwentyDashOne

\u{33f4}: '㏴'

IdeographicTelegraphSymbolForDayTwentyDashTwo

\u{33f5}: '㏵'

IdeographicTelegraphSymbolForDayTwentyDashThree

\u{33f6}: '㏶'

IdeographicTelegraphSymbolForDayTwentyDashFour

\u{33f7}: '㏷'

IdeographicTelegraphSymbolForDayTwentyDashFive

\u{33f8}: '㏸'

IdeographicTelegraphSymbolForDayTwentyDashSix

\u{33f9}: '㏹'

IdeographicTelegraphSymbolForDayTwentyDashSeven

\u{33fa}: '㏺'

IdeographicTelegraphSymbolForDayTwentyDashEight

\u{33fb}: '㏻'

IdeographicTelegraphSymbolForDayTwentyDashNine

\u{33fc}: '㏼'

IdeographicTelegraphSymbolForDayThirty

\u{33fd}: '㏽'

IdeographicTelegraphSymbolForDayThirtyDashOne

\u{33fe}: '㏾'

Methods

impl CJKCompatibility

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case