Enum unicode_types::cjk_compatibility_ideographs::CJKCompatibilityIdeographs

pub enum CJKCompatibilityIdeographs {
    CjkCompatibilityIdeographDashF900,
    CjkCompatibilityIdeographDashF901,
    CjkCompatibilityIdeographDashF902,
    CjkCompatibilityIdeographDashF903,
    CjkCompatibilityIdeographDashF904,
    CjkCompatibilityIdeographDashF905,
    CjkCompatibilityIdeographDashF906,
    CjkCompatibilityIdeographDashF907,
    CjkCompatibilityIdeographDashF908,
    CjkCompatibilityIdeographDashF909,
    CjkCompatibilityIdeographDashF90a,
    CjkCompatibilityIdeographDashF90b,
    CjkCompatibilityIdeographDashF90c,
    CjkCompatibilityIdeographDashF90d,
    CjkCompatibilityIdeographDashF90e,
    CjkCompatibilityIdeographDashF90f,
    CjkCompatibilityIdeographDashF910,
    CjkCompatibilityIdeographDashF911,
    CjkCompatibilityIdeographDashF912,
    CjkCompatibilityIdeographDashF913,
    CjkCompatibilityIdeographDashF914,
    CjkCompatibilityIdeographDashF915,
    CjkCompatibilityIdeographDashF916,
    CjkCompatibilityIdeographDashF917,
    CjkCompatibilityIdeographDashF918,
    CjkCompatibilityIdeographDashF919,
    CjkCompatibilityIdeographDashF91a,
    CjkCompatibilityIdeographDashF91b,
    CjkCompatibilityIdeographDashF91c,
    CjkCompatibilityIdeographDashF91d,
    CjkCompatibilityIdeographDashF91e,
    CjkCompatibilityIdeographDashF91f,
    CjkCompatibilityIdeographDashF920,
    CjkCompatibilityIdeographDashF921,
    CjkCompatibilityIdeographDashF922,
    CjkCompatibilityIdeographDashF923,
    CjkCompatibilityIdeographDashF924,
    CjkCompatibilityIdeographDashF925,
    CjkCompatibilityIdeographDashF926,
    CjkCompatibilityIdeographDashF927,
    CjkCompatibilityIdeographDashF928,
    CjkCompatibilityIdeographDashF929,
    CjkCompatibilityIdeographDashF92a,
    CjkCompatibilityIdeographDashF92b,
    CjkCompatibilityIdeographDashF92c,
    CjkCompatibilityIdeographDashF92d,
    CjkCompatibilityIdeographDashF92e,
    CjkCompatibilityIdeographDashF92f,
    CjkCompatibilityIdeographDashF930,
    CjkCompatibilityIdeographDashF931,
    CjkCompatibilityIdeographDashF932,
    CjkCompatibilityIdeographDashF933,
    CjkCompatibilityIdeographDashF934,
    CjkCompatibilityIdeographDashF935,
    CjkCompatibilityIdeographDashF936,
    CjkCompatibilityIdeographDashF937,
    CjkCompatibilityIdeographDashF938,
    CjkCompatibilityIdeographDashF939,
    CjkCompatibilityIdeographDashF93a,
    CjkCompatibilityIdeographDashF93b,
    CjkCompatibilityIdeographDashF93c,
    CjkCompatibilityIdeographDashF93d,
    CjkCompatibilityIdeographDashF93e,
    CjkCompatibilityIdeographDashF93f,
    CjkCompatibilityIdeographDashF940,
    CjkCompatibilityIdeographDashF941,
    CjkCompatibilityIdeographDashF942,
    CjkCompatibilityIdeographDashF943,
    CjkCompatibilityIdeographDashF944,
    CjkCompatibilityIdeographDashF945,
    CjkCompatibilityIdeographDashF946,
    CjkCompatibilityIdeographDashF947,
    CjkCompatibilityIdeographDashF948,
    CjkCompatibilityIdeographDashF949,
    CjkCompatibilityIdeographDashF94a,
    CjkCompatibilityIdeographDashF94b,
    CjkCompatibilityIdeographDashF94c,
    CjkCompatibilityIdeographDashF94d,
    CjkCompatibilityIdeographDashF94e,
    CjkCompatibilityIdeographDashF94f,
    CjkCompatibilityIdeographDashF950,
    CjkCompatibilityIdeographDashF951,
    CjkCompatibilityIdeographDashF952,
    CjkCompatibilityIdeographDashF953,
    CjkCompatibilityIdeographDashF954,
    CjkCompatibilityIdeographDashF955,
    CjkCompatibilityIdeographDashF956,
    CjkCompatibilityIdeographDashF957,
    CjkCompatibilityIdeographDashF958,
    CjkCompatibilityIdeographDashF959,
    CjkCompatibilityIdeographDashF95a,
    CjkCompatibilityIdeographDashF95b,
    CjkCompatibilityIdeographDashF95c,
    CjkCompatibilityIdeographDashF95d,
    CjkCompatibilityIdeographDashF95e,
    CjkCompatibilityIdeographDashF95f,
    CjkCompatibilityIdeographDashF960,
    CjkCompatibilityIdeographDashF961,
    CjkCompatibilityIdeographDashF962,
    CjkCompatibilityIdeographDashF963,
    CjkCompatibilityIdeographDashF964,
    CjkCompatibilityIdeographDashF965,
    CjkCompatibilityIdeographDashF966,
    CjkCompatibilityIdeographDashF967,
    CjkCompatibilityIdeographDashF968,
    CjkCompatibilityIdeographDashF969,
    CjkCompatibilityIdeographDashF96a,
    CjkCompatibilityIdeographDashF96b,
    CjkCompatibilityIdeographDashF96c,
    CjkCompatibilityIdeographDashF96d,
    CjkCompatibilityIdeographDashF96e,
    CjkCompatibilityIdeographDashF96f,
    CjkCompatibilityIdeographDashF970,
    CjkCompatibilityIdeographDashF971,
    CjkCompatibilityIdeographDashF972,
    CjkCompatibilityIdeographDashF973,
    CjkCompatibilityIdeographDashF974,
    CjkCompatibilityIdeographDashF975,
    CjkCompatibilityIdeographDashF976,
    CjkCompatibilityIdeographDashF977,
    CjkCompatibilityIdeographDashF978,
    CjkCompatibilityIdeographDashF979,
    CjkCompatibilityIdeographDashF97a,
    CjkCompatibilityIdeographDashF97b,
    CjkCompatibilityIdeographDashF97c,
    CjkCompatibilityIdeographDashF97d,
    CjkCompatibilityIdeographDashF97e,
    CjkCompatibilityIdeographDashF97f,
    CjkCompatibilityIdeographDashF980,
    CjkCompatibilityIdeographDashF981,
    CjkCompatibilityIdeographDashF982,
    CjkCompatibilityIdeographDashF983,
    CjkCompatibilityIdeographDashF984,
    CjkCompatibilityIdeographDashF985,
    CjkCompatibilityIdeographDashF986,
    CjkCompatibilityIdeographDashF987,
    CjkCompatibilityIdeographDashF988,
    CjkCompatibilityIdeographDashF989,
    CjkCompatibilityIdeographDashF98a,
    CjkCompatibilityIdeographDashF98b,
    CjkCompatibilityIdeographDashF98c,
    CjkCompatibilityIdeographDashF98d,
    CjkCompatibilityIdeographDashF98e,
    CjkCompatibilityIdeographDashF98f,
    CjkCompatibilityIdeographDashF990,
    CjkCompatibilityIdeographDashF991,
    CjkCompatibilityIdeographDashF992,
    CjkCompatibilityIdeographDashF993,
    CjkCompatibilityIdeographDashF994,
    CjkCompatibilityIdeographDashF995,
    CjkCompatibilityIdeographDashF996,
    CjkCompatibilityIdeographDashF997,
    CjkCompatibilityIdeographDashF998,
    CjkCompatibilityIdeographDashF999,
    CjkCompatibilityIdeographDashF99a,
    CjkCompatibilityIdeographDashF99b,
    CjkCompatibilityIdeographDashF99c,
    CjkCompatibilityIdeographDashF99d,
    CjkCompatibilityIdeographDashF99e,
    CjkCompatibilityIdeographDashF99f,
    CjkCompatibilityIdeographDashF9a0,
    CjkCompatibilityIdeographDashF9a1,
    CjkCompatibilityIdeographDashF9a2,
    CjkCompatibilityIdeographDashF9a3,
    CjkCompatibilityIdeographDashF9a4,
    CjkCompatibilityIdeographDashF9a5,
    CjkCompatibilityIdeographDashF9a6,
    CjkCompatibilityIdeographDashF9a7,
    CjkCompatibilityIdeographDashF9a8,
    CjkCompatibilityIdeographDashF9a9,
    CjkCompatibilityIdeographDashF9aa,
    CjkCompatibilityIdeographDashF9ab,
    CjkCompatibilityIdeographDashF9ac,
    CjkCompatibilityIdeographDashF9ad,
    CjkCompatibilityIdeographDashF9ae,
    CjkCompatibilityIdeographDashF9af,
    CjkCompatibilityIdeographDashF9b0,
    CjkCompatibilityIdeographDashF9b1,
    CjkCompatibilityIdeographDashF9b2,
    CjkCompatibilityIdeographDashF9b3,
    CjkCompatibilityIdeographDashF9b4,
    CjkCompatibilityIdeographDashF9b5,
    CjkCompatibilityIdeographDashF9b6,
    CjkCompatibilityIdeographDashF9b7,
    CjkCompatibilityIdeographDashF9b8,
    CjkCompatibilityIdeographDashF9b9,
    CjkCompatibilityIdeographDashF9ba,
    CjkCompatibilityIdeographDashF9bb,
    CjkCompatibilityIdeographDashF9bc,
    CjkCompatibilityIdeographDashF9bd,
    CjkCompatibilityIdeographDashF9be,
    CjkCompatibilityIdeographDashF9bf,
    CjkCompatibilityIdeographDashF9c0,
    CjkCompatibilityIdeographDashF9c1,
    CjkCompatibilityIdeographDashF9c2,
    CjkCompatibilityIdeographDashF9c3,
    CjkCompatibilityIdeographDashF9c4,
    CjkCompatibilityIdeographDashF9c5,
    CjkCompatibilityIdeographDashF9c6,
    CjkCompatibilityIdeographDashF9c7,
    CjkCompatibilityIdeographDashF9c8,
    CjkCompatibilityIdeographDashF9c9,
    CjkCompatibilityIdeographDashF9ca,
    CjkCompatibilityIdeographDashF9cb,
    CjkCompatibilityIdeographDashF9cc,
    CjkCompatibilityIdeographDashF9cd,
    CjkCompatibilityIdeographDashF9ce,
    CjkCompatibilityIdeographDashF9cf,
    CjkCompatibilityIdeographDashF9d0,
    CjkCompatibilityIdeographDashF9d1,
    CjkCompatibilityIdeographDashF9d2,
    CjkCompatibilityIdeographDashF9d3,
    CjkCompatibilityIdeographDashF9d4,
    CjkCompatibilityIdeographDashF9d5,
    CjkCompatibilityIdeographDashF9d6,
    CjkCompatibilityIdeographDashF9d7,
    CjkCompatibilityIdeographDashF9d8,
    CjkCompatibilityIdeographDashF9d9,
    CjkCompatibilityIdeographDashF9da,
    CjkCompatibilityIdeographDashF9db,
    CjkCompatibilityIdeographDashF9dc,
    CjkCompatibilityIdeographDashF9dd,
    CjkCompatibilityIdeographDashF9de,
    CjkCompatibilityIdeographDashF9df,
    CjkCompatibilityIdeographDashF9e0,
    CjkCompatibilityIdeographDashF9e1,
    CjkCompatibilityIdeographDashF9e2,
    CjkCompatibilityIdeographDashF9e3,
    CjkCompatibilityIdeographDashF9e4,
    CjkCompatibilityIdeographDashF9e5,
    CjkCompatibilityIdeographDashF9e6,
    CjkCompatibilityIdeographDashF9e7,
    CjkCompatibilityIdeographDashF9e8,
    CjkCompatibilityIdeographDashF9e9,
    CjkCompatibilityIdeographDashF9ea,
    CjkCompatibilityIdeographDashF9eb,
    CjkCompatibilityIdeographDashF9ec,
    CjkCompatibilityIdeographDashF9ed,
    CjkCompatibilityIdeographDashF9ee,
    CjkCompatibilityIdeographDashF9ef,
    CjkCompatibilityIdeographDashF9f0,
    CjkCompatibilityIdeographDashF9f1,
    CjkCompatibilityIdeographDashF9f2,
    CjkCompatibilityIdeographDashF9f3,
    CjkCompatibilityIdeographDashF9f4,
    CjkCompatibilityIdeographDashF9f5,
    CjkCompatibilityIdeographDashF9f6,
    CjkCompatibilityIdeographDashF9f7,
    CjkCompatibilityIdeographDashF9f8,
    CjkCompatibilityIdeographDashF9f9,
    CjkCompatibilityIdeographDashF9fa,
    CjkCompatibilityIdeographDashF9fb,
    CjkCompatibilityIdeographDashF9fc,
    CjkCompatibilityIdeographDashF9fd,
    CjkCompatibilityIdeographDashF9fe,
    CjkCompatibilityIdeographDashF9ff,
    CjkCompatibilityIdeographDashFa00,
    CjkCompatibilityIdeographDashFa01,
    CjkCompatibilityIdeographDashFa02,
    CjkCompatibilityIdeographDashFa03,
    CjkCompatibilityIdeographDashFa04,
    CjkCompatibilityIdeographDashFa05,
    CjkCompatibilityIdeographDashFa06,
    CjkCompatibilityIdeographDashFa07,
    CjkCompatibilityIdeographDashFa08,
    CjkCompatibilityIdeographDashFa09,
    CjkCompatibilityIdeographDashFa0a,
    CjkCompatibilityIdeographDashFa0b,
    CjkCompatibilityIdeographDashFa0c,
    CjkCompatibilityIdeographDashFa0d,
    CjkCompatibilityIdeographDashFa0e,
    CjkCompatibilityIdeographDashFa0f,
    CjkCompatibilityIdeographDashFa10,
    CjkCompatibilityIdeographDashFa11,
    CjkCompatibilityIdeographDashFa12,
    CjkCompatibilityIdeographDashFa13,
    CjkCompatibilityIdeographDashFa14,
    CjkCompatibilityIdeographDashFa15,
    CjkCompatibilityIdeographDashFa16,
    CjkCompatibilityIdeographDashFa17,
    CjkCompatibilityIdeographDashFa18,
    CjkCompatibilityIdeographDashFa19,
    CjkCompatibilityIdeographDashFa1a,
    CjkCompatibilityIdeographDashFa1b,
    CjkCompatibilityIdeographDashFa1c,
    CjkCompatibilityIdeographDashFa1d,
    CjkCompatibilityIdeographDashFa1e,
    CjkCompatibilityIdeographDashFa1f,
    CjkCompatibilityIdeographDashFa20,
    CjkCompatibilityIdeographDashFa21,
    CjkCompatibilityIdeographDashFa22,
    CjkCompatibilityIdeographDashFa23,
    CjkCompatibilityIdeographDashFa24,
    CjkCompatibilityIdeographDashFa25,
    CjkCompatibilityIdeographDashFa26,
    CjkCompatibilityIdeographDashFa27,
    CjkCompatibilityIdeographDashFa28,
    CjkCompatibilityIdeographDashFa29,
    CjkCompatibilityIdeographDashFa2a,
    CjkCompatibilityIdeographDashFa2b,
    CjkCompatibilityIdeographDashFa2c,
    CjkCompatibilityIdeographDashFa2d,
    CjkCompatibilityIdeographDashFa2e,
    CjkCompatibilityIdeographDashFa2f,
    CjkCompatibilityIdeographDashFa30,
    CjkCompatibilityIdeographDashFa31,
    CjkCompatibilityIdeographDashFa32,
    CjkCompatibilityIdeographDashFa33,
    CjkCompatibilityIdeographDashFa34,
    CjkCompatibilityIdeographDashFa35,
    CjkCompatibilityIdeographDashFa36,
    CjkCompatibilityIdeographDashFa37,
    CjkCompatibilityIdeographDashFa38,
    CjkCompatibilityIdeographDashFa39,
    CjkCompatibilityIdeographDashFa3a,
    CjkCompatibilityIdeographDashFa3b,
    CjkCompatibilityIdeographDashFa3c,
    CjkCompatibilityIdeographDashFa3d,
    CjkCompatibilityIdeographDashFa3e,
    CjkCompatibilityIdeographDashFa3f,
    CjkCompatibilityIdeographDashFa40,
    CjkCompatibilityIdeographDashFa41,
    CjkCompatibilityIdeographDashFa42,
    CjkCompatibilityIdeographDashFa43,
    CjkCompatibilityIdeographDashFa44,
    CjkCompatibilityIdeographDashFa45,
    CjkCompatibilityIdeographDashFa46,
    CjkCompatibilityIdeographDashFa47,
    CjkCompatibilityIdeographDashFa48,
    CjkCompatibilityIdeographDashFa49,
    CjkCompatibilityIdeographDashFa4a,
    CjkCompatibilityIdeographDashFa4b,
    CjkCompatibilityIdeographDashFa4c,
    CjkCompatibilityIdeographDashFa4d,
    CjkCompatibilityIdeographDashFa4e,
    CjkCompatibilityIdeographDashFa4f,
    CjkCompatibilityIdeographDashFa50,
    CjkCompatibilityIdeographDashFa51,
    CjkCompatibilityIdeographDashFa52,
    CjkCompatibilityIdeographDashFa53,
    CjkCompatibilityIdeographDashFa54,
    CjkCompatibilityIdeographDashFa55,
    CjkCompatibilityIdeographDashFa56,
    CjkCompatibilityIdeographDashFa57,
    CjkCompatibilityIdeographDashFa58,
    CjkCompatibilityIdeographDashFa59,
    CjkCompatibilityIdeographDashFa5a,
    CjkCompatibilityIdeographDashFa5b,
    CjkCompatibilityIdeographDashFa5c,
    CjkCompatibilityIdeographDashFa5d,
    CjkCompatibilityIdeographDashFa5e,
    CjkCompatibilityIdeographDashFa5f,
    CjkCompatibilityIdeographDashFa60,
    CjkCompatibilityIdeographDashFa61,
    CjkCompatibilityIdeographDashFa62,
    CjkCompatibilityIdeographDashFa63,
    CjkCompatibilityIdeographDashFa64,
    CjkCompatibilityIdeographDashFa65,
    CjkCompatibilityIdeographDashFa66,
    CjkCompatibilityIdeographDashFa67,
    CjkCompatibilityIdeographDashFa68,
    CjkCompatibilityIdeographDashFa69,
    CjkCompatibilityIdeographDashFa6a,
    CjkCompatibilityIdeographDashFa6b,
    CjkCompatibilityIdeographDashFa6c,
    CjkCompatibilityIdeographDashFa6d,
    CjkCompatibilityIdeographDashFa70,
    CjkCompatibilityIdeographDashFa71,
    CjkCompatibilityIdeographDashFa72,
    CjkCompatibilityIdeographDashFa73,
    CjkCompatibilityIdeographDashFa74,
    CjkCompatibilityIdeographDashFa75,
    CjkCompatibilityIdeographDashFa76,
    CjkCompatibilityIdeographDashFa77,
    CjkCompatibilityIdeographDashFa78,
    CjkCompatibilityIdeographDashFa79,
    CjkCompatibilityIdeographDashFa7a,
    CjkCompatibilityIdeographDashFa7b,
    CjkCompatibilityIdeographDashFa7c,
    CjkCompatibilityIdeographDashFa7d,
    CjkCompatibilityIdeographDashFa7e,
    CjkCompatibilityIdeographDashFa7f,
    CjkCompatibilityIdeographDashFa80,
    CjkCompatibilityIdeographDashFa81,
    CjkCompatibilityIdeographDashFa82,
    CjkCompatibilityIdeographDashFa83,
    CjkCompatibilityIdeographDashFa84,
    CjkCompatibilityIdeographDashFa85,
    CjkCompatibilityIdeographDashFa86,
    CjkCompatibilityIdeographDashFa87,
    CjkCompatibilityIdeographDashFa88,
    CjkCompatibilityIdeographDashFa89,
    CjkCompatibilityIdeographDashFa8a,
    CjkCompatibilityIdeographDashFa8b,
    CjkCompatibilityIdeographDashFa8c,
    CjkCompatibilityIdeographDashFa8d,
    CjkCompatibilityIdeographDashFa8e,
    CjkCompatibilityIdeographDashFa8f,
    CjkCompatibilityIdeographDashFa90,
    CjkCompatibilityIdeographDashFa91,
    CjkCompatibilityIdeographDashFa92,
    CjkCompatibilityIdeographDashFa93,
    CjkCompatibilityIdeographDashFa94,
    CjkCompatibilityIdeographDashFa95,
    CjkCompatibilityIdeographDashFa96,
    CjkCompatibilityIdeographDashFa97,
    CjkCompatibilityIdeographDashFa98,
    CjkCompatibilityIdeographDashFa99,
    CjkCompatibilityIdeographDashFa9a,
    CjkCompatibilityIdeographDashFa9b,
    CjkCompatibilityIdeographDashFa9c,
    CjkCompatibilityIdeographDashFa9d,
    CjkCompatibilityIdeographDashFa9e,
    CjkCompatibilityIdeographDashFa9f,
    CjkCompatibilityIdeographDashFaa0,
    CjkCompatibilityIdeographDashFaa1,
    CjkCompatibilityIdeographDashFaa2,
    CjkCompatibilityIdeographDashFaa3,
    CjkCompatibilityIdeographDashFaa4,
    CjkCompatibilityIdeographDashFaa5,
    CjkCompatibilityIdeographDashFaa6,
    CjkCompatibilityIdeographDashFaa7,
    CjkCompatibilityIdeographDashFaa8,
    CjkCompatibilityIdeographDashFaa9,
    CjkCompatibilityIdeographDashFaaa,
    CjkCompatibilityIdeographDashFaab,
    CjkCompatibilityIdeographDashFaac,
    CjkCompatibilityIdeographDashFaad,
    CjkCompatibilityIdeographDashFaae,
    CjkCompatibilityIdeographDashFaaf,
    CjkCompatibilityIdeographDashFab0,
    CjkCompatibilityIdeographDashFab1,
    CjkCompatibilityIdeographDashFab2,
    CjkCompatibilityIdeographDashFab3,
    CjkCompatibilityIdeographDashFab4,
    CjkCompatibilityIdeographDashFab5,
    CjkCompatibilityIdeographDashFab6,
    CjkCompatibilityIdeographDashFab7,
    CjkCompatibilityIdeographDashFab8,
    CjkCompatibilityIdeographDashFab9,
    CjkCompatibilityIdeographDashFaba,
    CjkCompatibilityIdeographDashFabb,
    CjkCompatibilityIdeographDashFabc,
    CjkCompatibilityIdeographDashFabd,
    CjkCompatibilityIdeographDashFabe,
    CjkCompatibilityIdeographDashFabf,
    CjkCompatibilityIdeographDashFac0,
    CjkCompatibilityIdeographDashFac1,
    CjkCompatibilityIdeographDashFac2,
    CjkCompatibilityIdeographDashFac3,
    CjkCompatibilityIdeographDashFac4,
    CjkCompatibilityIdeographDashFac5,
    CjkCompatibilityIdeographDashFac6,
    CjkCompatibilityIdeographDashFac7,
    CjkCompatibilityIdeographDashFac8,
    CjkCompatibilityIdeographDashFac9,
    CjkCompatibilityIdeographDashFaca,
    CjkCompatibilityIdeographDashFacb,
    CjkCompatibilityIdeographDashFacc,
    CjkCompatibilityIdeographDashFacd,
    CjkCompatibilityIdeographDashFace,
    CjkCompatibilityIdeographDashFacf,
    CjkCompatibilityIdeographDashFad0,
    CjkCompatibilityIdeographDashFad1,
    CjkCompatibilityIdeographDashFad2,
    CjkCompatibilityIdeographDashFad3,
    CjkCompatibilityIdeographDashFad4,
    CjkCompatibilityIdeographDashFad5,
    CjkCompatibilityIdeographDashFad6,
    CjkCompatibilityIdeographDashFad7,
    CjkCompatibilityIdeographDashFad8,
    CjkCompatibilityIdeographDashFad9,
}

An enum to represent all characters in the CJKCompatibilityIdeographs block.

Variants
CjkCompatibilityIdeographDashF900

\u{f900}: '豈'

CjkCompatibilityIdeographDashF901

\u{f901}: '更'

CjkCompatibilityIdeographDashF902

\u{f902}: '車'

CjkCompatibilityIdeographDashF903

\u{f903}: '賈'

CjkCompatibilityIdeographDashF904

\u{f904}: '滑'

CjkCompatibilityIdeographDashF905

\u{f905}: '串'

CjkCompatibilityIdeographDashF906

\u{f906}: '句'

CjkCompatibilityIdeographDashF907

\u{f907}: '龜'

CjkCompatibilityIdeographDashF908

\u{f908}: '龜'

CjkCompatibilityIdeographDashF909

\u{f909}: '契'

CjkCompatibilityIdeographDashF90a

\u{f90a}: '金'

CjkCompatibilityIdeographDashF90b

\u{f90b}: '喇'

CjkCompatibilityIdeographDashF90c

\u{f90c}: '奈'

CjkCompatibilityIdeographDashF90d

\u{f90d}: '懶'

CjkCompatibilityIdeographDashF90e

\u{f90e}: '癩'

CjkCompatibilityIdeographDashF90f

\u{f90f}: '羅'

CjkCompatibilityIdeographDashF910

\u{f910}: '蘿'

CjkCompatibilityIdeographDashF911

\u{f911}: '螺'

CjkCompatibilityIdeographDashF912

\u{f912}: '裸'

CjkCompatibilityIdeographDashF913

\u{f913}: '邏'

CjkCompatibilityIdeographDashF914

\u{f914}: '樂'

CjkCompatibilityIdeographDashF915

\u{f915}: '洛'

CjkCompatibilityIdeographDashF916

\u{f916}: '烙'

CjkCompatibilityIdeographDashF917

\u{f917}: '珞'

CjkCompatibilityIdeographDashF918

\u{f918}: '落'

CjkCompatibilityIdeographDashF919

\u{f919}: '酪'

CjkCompatibilityIdeographDashF91a

\u{f91a}: '駱'

CjkCompatibilityIdeographDashF91b

\u{f91b}: '亂'

CjkCompatibilityIdeographDashF91c

\u{f91c}: '卵'

CjkCompatibilityIdeographDashF91d

\u{f91d}: '欄'

CjkCompatibilityIdeographDashF91e

\u{f91e}: '爛'

CjkCompatibilityIdeographDashF91f

\u{f91f}: '蘭'

CjkCompatibilityIdeographDashF920

\u{f920}: '鸞'

CjkCompatibilityIdeographDashF921

\u{f921}: '嵐'

CjkCompatibilityIdeographDashF922

\u{f922}: '濫'

CjkCompatibilityIdeographDashF923

\u{f923}: '藍'

CjkCompatibilityIdeographDashF924

\u{f924}: '襤'

CjkCompatibilityIdeographDashF925

\u{f925}: '拉'

CjkCompatibilityIdeographDashF926

\u{f926}: '臘'

CjkCompatibilityIdeographDashF927

\u{f927}: '蠟'

CjkCompatibilityIdeographDashF928

\u{f928}: '廊'

CjkCompatibilityIdeographDashF929

\u{f929}: '朗'

CjkCompatibilityIdeographDashF92a

\u{f92a}: '浪'

CjkCompatibilityIdeographDashF92b

\u{f92b}: '狼'

CjkCompatibilityIdeographDashF92c

\u{f92c}: '郎'

CjkCompatibilityIdeographDashF92d

\u{f92d}: '來'

CjkCompatibilityIdeographDashF92e

\u{f92e}: '冷'

CjkCompatibilityIdeographDashF92f

\u{f92f}: '勞'

CjkCompatibilityIdeographDashF930

\u{f930}: '擄'

CjkCompatibilityIdeographDashF931

\u{f931}: '櫓'

CjkCompatibilityIdeographDashF932

\u{f932}: '爐'

CjkCompatibilityIdeographDashF933

\u{f933}: '盧'

CjkCompatibilityIdeographDashF934

\u{f934}: '老'

CjkCompatibilityIdeographDashF935

\u{f935}: '蘆'

CjkCompatibilityIdeographDashF936

\u{f936}: '虜'

CjkCompatibilityIdeographDashF937

\u{f937}: '路'

CjkCompatibilityIdeographDashF938

\u{f938}: '露'

CjkCompatibilityIdeographDashF939

\u{f939}: '魯'

CjkCompatibilityIdeographDashF93a

\u{f93a}: '鷺'

CjkCompatibilityIdeographDashF93b

\u{f93b}: '碌'

CjkCompatibilityIdeographDashF93c

\u{f93c}: '祿'

CjkCompatibilityIdeographDashF93d

\u{f93d}: '綠'

CjkCompatibilityIdeographDashF93e

\u{f93e}: '菉'

CjkCompatibilityIdeographDashF93f

\u{f93f}: '錄'

CjkCompatibilityIdeographDashF940

\u{f940}: '鹿'

CjkCompatibilityIdeographDashF941

\u{f941}: '論'

CjkCompatibilityIdeographDashF942

\u{f942}: '壟'

CjkCompatibilityIdeographDashF943

\u{f943}: '弄'

CjkCompatibilityIdeographDashF944

\u{f944}: '籠'

CjkCompatibilityIdeographDashF945

\u{f945}: '聾'

CjkCompatibilityIdeographDashF946

\u{f946}: '牢'

CjkCompatibilityIdeographDashF947

\u{f947}: '磊'

CjkCompatibilityIdeographDashF948

\u{f948}: '賂'

CjkCompatibilityIdeographDashF949

\u{f949}: '雷'

CjkCompatibilityIdeographDashF94a

\u{f94a}: '壘'

CjkCompatibilityIdeographDashF94b

\u{f94b}: '屢'

CjkCompatibilityIdeographDashF94c

\u{f94c}: '樓'

CjkCompatibilityIdeographDashF94d

\u{f94d}: '淚'

CjkCompatibilityIdeographDashF94e

\u{f94e}: '漏'

CjkCompatibilityIdeographDashF94f

\u{f94f}: '累'

CjkCompatibilityIdeographDashF950

\u{f950}: '縷'

CjkCompatibilityIdeographDashF951

\u{f951}: '陋'

CjkCompatibilityIdeographDashF952

\u{f952}: '勒'

CjkCompatibilityIdeographDashF953

\u{f953}: '肋'

CjkCompatibilityIdeographDashF954

\u{f954}: '凜'

CjkCompatibilityIdeographDashF955

\u{f955}: '凌'

CjkCompatibilityIdeographDashF956

\u{f956}: '稜'

CjkCompatibilityIdeographDashF957

\u{f957}: '綾'

CjkCompatibilityIdeographDashF958

\u{f958}: '菱'

CjkCompatibilityIdeographDashF959

\u{f959}: '陵'

CjkCompatibilityIdeographDashF95a

\u{f95a}: '讀'

CjkCompatibilityIdeographDashF95b

\u{f95b}: '拏'

CjkCompatibilityIdeographDashF95c

\u{f95c}: '樂'

CjkCompatibilityIdeographDashF95d

\u{f95d}: '諾'

CjkCompatibilityIdeographDashF95e

\u{f95e}: '丹'

CjkCompatibilityIdeographDashF95f

\u{f95f}: '寧'

CjkCompatibilityIdeographDashF960

\u{f960}: '怒'

CjkCompatibilityIdeographDashF961

\u{f961}: '率'

CjkCompatibilityIdeographDashF962

\u{f962}: '異'

CjkCompatibilityIdeographDashF963

\u{f963}: '北'

CjkCompatibilityIdeographDashF964

\u{f964}: '磻'

CjkCompatibilityIdeographDashF965

\u{f965}: '便'

CjkCompatibilityIdeographDashF966

\u{f966}: '復'

CjkCompatibilityIdeographDashF967

\u{f967}: '不'

CjkCompatibilityIdeographDashF968

\u{f968}: '泌'

CjkCompatibilityIdeographDashF969

\u{f969}: '數'

CjkCompatibilityIdeographDashF96a

\u{f96a}: '索'

CjkCompatibilityIdeographDashF96b

\u{f96b}: '參'

CjkCompatibilityIdeographDashF96c

\u{f96c}: '塞'

CjkCompatibilityIdeographDashF96d

\u{f96d}: '省'

CjkCompatibilityIdeographDashF96e

\u{f96e}: '葉'

CjkCompatibilityIdeographDashF96f

\u{f96f}: '說'

CjkCompatibilityIdeographDashF970

\u{f970}: '殺'

CjkCompatibilityIdeographDashF971

\u{f971}: '辰'

CjkCompatibilityIdeographDashF972

\u{f972}: '沈'

CjkCompatibilityIdeographDashF973

\u{f973}: '拾'

CjkCompatibilityIdeographDashF974

\u{f974}: '若'

CjkCompatibilityIdeographDashF975

\u{f975}: '掠'

CjkCompatibilityIdeographDashF976

\u{f976}: '略'

CjkCompatibilityIdeographDashF977

\u{f977}: '亮'

CjkCompatibilityIdeographDashF978

\u{f978}: '兩'

CjkCompatibilityIdeographDashF979

\u{f979}: '凉'

CjkCompatibilityIdeographDashF97a

\u{f97a}: '梁'

CjkCompatibilityIdeographDashF97b

\u{f97b}: '糧'

CjkCompatibilityIdeographDashF97c

\u{f97c}: '良'

CjkCompatibilityIdeographDashF97d

\u{f97d}: '諒'

CjkCompatibilityIdeographDashF97e

\u{f97e}: '量'

CjkCompatibilityIdeographDashF97f

\u{f97f}: '勵'

CjkCompatibilityIdeographDashF980

\u{f980}: '呂'

CjkCompatibilityIdeographDashF981

\u{f981}: '女'

CjkCompatibilityIdeographDashF982

\u{f982}: '廬'

CjkCompatibilityIdeographDashF983

\u{f983}: '旅'

CjkCompatibilityIdeographDashF984

\u{f984}: '濾'

CjkCompatibilityIdeographDashF985

\u{f985}: '礪'

CjkCompatibilityIdeographDashF986

\u{f986}: '閭'

CjkCompatibilityIdeographDashF987

\u{f987}: '驪'

CjkCompatibilityIdeographDashF988

\u{f988}: '麗'

CjkCompatibilityIdeographDashF989

\u{f989}: '黎'

CjkCompatibilityIdeographDashF98a

\u{f98a}: '力'

CjkCompatibilityIdeographDashF98b

\u{f98b}: '曆'

CjkCompatibilityIdeographDashF98c

\u{f98c}: '歷'

CjkCompatibilityIdeographDashF98d

\u{f98d}: '轢'

CjkCompatibilityIdeographDashF98e

\u{f98e}: '年'

CjkCompatibilityIdeographDashF98f

\u{f98f}: '憐'

CjkCompatibilityIdeographDashF990

\u{f990}: '戀'

CjkCompatibilityIdeographDashF991

\u{f991}: '撚'

CjkCompatibilityIdeographDashF992

\u{f992}: '漣'

CjkCompatibilityIdeographDashF993

\u{f993}: '煉'

CjkCompatibilityIdeographDashF994

\u{f994}: '璉'

CjkCompatibilityIdeographDashF995

\u{f995}: '秊'

CjkCompatibilityIdeographDashF996

\u{f996}: '練'

CjkCompatibilityIdeographDashF997

\u{f997}: '聯'

CjkCompatibilityIdeographDashF998

\u{f998}: '輦'

CjkCompatibilityIdeographDashF999

\u{f999}: '蓮'

CjkCompatibilityIdeographDashF99a

\u{f99a}: '連'

CjkCompatibilityIdeographDashF99b

\u{f99b}: '鍊'

CjkCompatibilityIdeographDashF99c

\u{f99c}: '列'

CjkCompatibilityIdeographDashF99d

\u{f99d}: '劣'

CjkCompatibilityIdeographDashF99e

\u{f99e}: '咽'

CjkCompatibilityIdeographDashF99f

\u{f99f}: '烈'

CjkCompatibilityIdeographDashF9a0

\u{f9a0}: '裂'

CjkCompatibilityIdeographDashF9a1

\u{f9a1}: '說'

CjkCompatibilityIdeographDashF9a2

\u{f9a2}: '廉'

CjkCompatibilityIdeographDashF9a3

\u{f9a3}: '念'

CjkCompatibilityIdeographDashF9a4

\u{f9a4}: '捻'

CjkCompatibilityIdeographDashF9a5

\u{f9a5}: '殮'

CjkCompatibilityIdeographDashF9a6

\u{f9a6}: '簾'

CjkCompatibilityIdeographDashF9a7

\u{f9a7}: '獵'

CjkCompatibilityIdeographDashF9a8

\u{f9a8}: '令'

CjkCompatibilityIdeographDashF9a9

\u{f9a9}: '囹'

CjkCompatibilityIdeographDashF9aa

\u{f9aa}: '寧'

CjkCompatibilityIdeographDashF9ab

\u{f9ab}: '嶺'

CjkCompatibilityIdeographDashF9ac

\u{f9ac}: '怜'

CjkCompatibilityIdeographDashF9ad

\u{f9ad}: '玲'

CjkCompatibilityIdeographDashF9ae

\u{f9ae}: '瑩'

CjkCompatibilityIdeographDashF9af

\u{f9af}: '羚'

CjkCompatibilityIdeographDashF9b0

\u{f9b0}: '聆'

CjkCompatibilityIdeographDashF9b1

\u{f9b1}: '鈴'

CjkCompatibilityIdeographDashF9b2

\u{f9b2}: '零'

CjkCompatibilityIdeographDashF9b3

\u{f9b3}: '靈'

CjkCompatibilityIdeographDashF9b4

\u{f9b4}: '領'

CjkCompatibilityIdeographDashF9b5

\u{f9b5}: '例'

CjkCompatibilityIdeographDashF9b6

\u{f9b6}: '禮'

CjkCompatibilityIdeographDashF9b7

\u{f9b7}: '醴'

CjkCompatibilityIdeographDashF9b8

\u{f9b8}: '隸'

CjkCompatibilityIdeographDashF9b9

\u{f9b9}: '惡'

CjkCompatibilityIdeographDashF9ba

\u{f9ba}: '了'

CjkCompatibilityIdeographDashF9bb

\u{f9bb}: '僚'

CjkCompatibilityIdeographDashF9bc

\u{f9bc}: '寮'

CjkCompatibilityIdeographDashF9bd

\u{f9bd}: '尿'

CjkCompatibilityIdeographDashF9be

\u{f9be}: '料'

CjkCompatibilityIdeographDashF9bf

\u{f9bf}: '樂'

CjkCompatibilityIdeographDashF9c0

\u{f9c0}: '燎'

CjkCompatibilityIdeographDashF9c1

\u{f9c1}: '療'

CjkCompatibilityIdeographDashF9c2

\u{f9c2}: '蓼'

CjkCompatibilityIdeographDashF9c3

\u{f9c3}: '遼'

CjkCompatibilityIdeographDashF9c4

\u{f9c4}: '龍'

CjkCompatibilityIdeographDashF9c5

\u{f9c5}: '暈'

CjkCompatibilityIdeographDashF9c6

\u{f9c6}: '阮'

CjkCompatibilityIdeographDashF9c7

\u{f9c7}: '劉'

CjkCompatibilityIdeographDashF9c8

\u{f9c8}: '杻'

CjkCompatibilityIdeographDashF9c9

\u{f9c9}: '柳'

CjkCompatibilityIdeographDashF9ca

\u{f9ca}: '流'

CjkCompatibilityIdeographDashF9cb

\u{f9cb}: '溜'

CjkCompatibilityIdeographDashF9cc

\u{f9cc}: '琉'

CjkCompatibilityIdeographDashF9cd

\u{f9cd}: '留'

CjkCompatibilityIdeographDashF9ce

\u{f9ce}: '硫'

CjkCompatibilityIdeographDashF9cf

\u{f9cf}: '紐'

CjkCompatibilityIdeographDashF9d0

\u{f9d0}: '類'

CjkCompatibilityIdeographDashF9d1

\u{f9d1}: '六'

CjkCompatibilityIdeographDashF9d2

\u{f9d2}: '戮'

CjkCompatibilityIdeographDashF9d3

\u{f9d3}: '陸'

CjkCompatibilityIdeographDashF9d4

\u{f9d4}: '倫'

CjkCompatibilityIdeographDashF9d5

\u{f9d5}: '崙'

CjkCompatibilityIdeographDashF9d6

\u{f9d6}: '淪'

CjkCompatibilityIdeographDashF9d7

\u{f9d7}: '輪'

CjkCompatibilityIdeographDashF9d8

\u{f9d8}: '律'

CjkCompatibilityIdeographDashF9d9

\u{f9d9}: '慄'

CjkCompatibilityIdeographDashF9da

\u{f9da}: '栗'

CjkCompatibilityIdeographDashF9db

\u{f9db}: '率'

CjkCompatibilityIdeographDashF9dc

\u{f9dc}: '隆'

CjkCompatibilityIdeographDashF9dd

\u{f9dd}: '利'

CjkCompatibilityIdeographDashF9de

\u{f9de}: '吏'

CjkCompatibilityIdeographDashF9df

\u{f9df}: '履'

CjkCompatibilityIdeographDashF9e0

\u{f9e0}: '易'

CjkCompatibilityIdeographDashF9e1

\u{f9e1}: '李'

CjkCompatibilityIdeographDashF9e2

\u{f9e2}: '梨'

CjkCompatibilityIdeographDashF9e3

\u{f9e3}: '泥'

CjkCompatibilityIdeographDashF9e4

\u{f9e4}: '理'

CjkCompatibilityIdeographDashF9e5

\u{f9e5}: '痢'

CjkCompatibilityIdeographDashF9e6

\u{f9e6}: '罹'

CjkCompatibilityIdeographDashF9e7

\u{f9e7}: '裏'

CjkCompatibilityIdeographDashF9e8

\u{f9e8}: '裡'

CjkCompatibilityIdeographDashF9e9

\u{f9e9}: '里'

CjkCompatibilityIdeographDashF9ea

\u{f9ea}: '離'

CjkCompatibilityIdeographDashF9eb

\u{f9eb}: '匿'

CjkCompatibilityIdeographDashF9ec

\u{f9ec}: '溺'

CjkCompatibilityIdeographDashF9ed

\u{f9ed}: '吝'

CjkCompatibilityIdeographDashF9ee

\u{f9ee}: '燐'

CjkCompatibilityIdeographDashF9ef

\u{f9ef}: '璘'

CjkCompatibilityIdeographDashF9f0

\u{f9f0}: '藺'

CjkCompatibilityIdeographDashF9f1

\u{f9f1}: '隣'

CjkCompatibilityIdeographDashF9f2

\u{f9f2}: '鱗'

CjkCompatibilityIdeographDashF9f3

\u{f9f3}: '麟'

CjkCompatibilityIdeographDashF9f4

\u{f9f4}: '林'

CjkCompatibilityIdeographDashF9f5

\u{f9f5}: '淋'

CjkCompatibilityIdeographDashF9f6

\u{f9f6}: '臨'

CjkCompatibilityIdeographDashF9f7

\u{f9f7}: '立'

CjkCompatibilityIdeographDashF9f8

\u{f9f8}: '笠'

CjkCompatibilityIdeographDashF9f9

\u{f9f9}: '粒'

CjkCompatibilityIdeographDashF9fa

\u{f9fa}: '狀'

CjkCompatibilityIdeographDashF9fb

\u{f9fb}: '炙'

CjkCompatibilityIdeographDashF9fc

\u{f9fc}: '識'

CjkCompatibilityIdeographDashF9fd

\u{f9fd}: '什'

CjkCompatibilityIdeographDashF9fe

\u{f9fe}: '茶'

CjkCompatibilityIdeographDashF9ff

\u{f9ff}: '刺'

CjkCompatibilityIdeographDashFa00

\u{fa00}: '切'

CjkCompatibilityIdeographDashFa01

\u{fa01}: '度'

CjkCompatibilityIdeographDashFa02

\u{fa02}: '拓'

CjkCompatibilityIdeographDashFa03

\u{fa03}: '糖'

CjkCompatibilityIdeographDashFa04

\u{fa04}: '宅'

CjkCompatibilityIdeographDashFa05

\u{fa05}: '洞'

CjkCompatibilityIdeographDashFa06

\u{fa06}: '暴'

CjkCompatibilityIdeographDashFa07

\u{fa07}: '輻'

CjkCompatibilityIdeographDashFa08

\u{fa08}: '行'

CjkCompatibilityIdeographDashFa09

\u{fa09}: '降'

CjkCompatibilityIdeographDashFa0a

\u{fa0a}: '見'

CjkCompatibilityIdeographDashFa0b

\u{fa0b}: '廓'

CjkCompatibilityIdeographDashFa0c

\u{fa0c}: '兀'

CjkCompatibilityIdeographDashFa0d

\u{fa0d}: '嗀'

CjkCompatibilityIdeographDashFa0e

\u{fa0e}: '﨎'

CjkCompatibilityIdeographDashFa0f

\u{fa0f}: '﨏'

CjkCompatibilityIdeographDashFa10

\u{fa10}: '塚'

CjkCompatibilityIdeographDashFa11

\u{fa11}: '﨑'

CjkCompatibilityIdeographDashFa12

\u{fa12}: '晴'

CjkCompatibilityIdeographDashFa13

\u{fa13}: '﨓'

CjkCompatibilityIdeographDashFa14

\u{fa14}: '﨔'

CjkCompatibilityIdeographDashFa15

\u{fa15}: '凞'

CjkCompatibilityIdeographDashFa16

\u{fa16}: '猪'

CjkCompatibilityIdeographDashFa17

\u{fa17}: '益'

CjkCompatibilityIdeographDashFa18

\u{fa18}: '礼'

CjkCompatibilityIdeographDashFa19

\u{fa19}: '神'

CjkCompatibilityIdeographDashFa1a

\u{fa1a}: '祥'

CjkCompatibilityIdeographDashFa1b

\u{fa1b}: '福'

CjkCompatibilityIdeographDashFa1c

\u{fa1c}: '靖'

CjkCompatibilityIdeographDashFa1d

\u{fa1d}: '精'

CjkCompatibilityIdeographDashFa1e

\u{fa1e}: '羽'

CjkCompatibilityIdeographDashFa1f

\u{fa1f}: '﨟'

CjkCompatibilityIdeographDashFa20

\u{fa20}: '蘒'

CjkCompatibilityIdeographDashFa21

\u{fa21}: '﨡'

CjkCompatibilityIdeographDashFa22

\u{fa22}: '諸'

CjkCompatibilityIdeographDashFa23

\u{fa23}: '﨣'

CjkCompatibilityIdeographDashFa24

\u{fa24}: '﨤'

CjkCompatibilityIdeographDashFa25

\u{fa25}: '逸'

CjkCompatibilityIdeographDashFa26

\u{fa26}: '都'

CjkCompatibilityIdeographDashFa27

\u{fa27}: '﨧'

CjkCompatibilityIdeographDashFa28

\u{fa28}: '﨨'

CjkCompatibilityIdeographDashFa29

\u{fa29}: '﨩'

CjkCompatibilityIdeographDashFa2a

\u{fa2a}: '飯'

CjkCompatibilityIdeographDashFa2b

\u{fa2b}: '飼'

CjkCompatibilityIdeographDashFa2c

\u{fa2c}: '館'

CjkCompatibilityIdeographDashFa2d

\u{fa2d}: '鶴'

CjkCompatibilityIdeographDashFa2e

\u{fa2e}: '郞'

CjkCompatibilityIdeographDashFa2f

\u{fa2f}: '隷'

CjkCompatibilityIdeographDashFa30

\u{fa30}: '侮'

CjkCompatibilityIdeographDashFa31

\u{fa31}: '僧'

CjkCompatibilityIdeographDashFa32

\u{fa32}: '免'

CjkCompatibilityIdeographDashFa33

\u{fa33}: '勉'

CjkCompatibilityIdeographDashFa34

\u{fa34}: '勤'

CjkCompatibilityIdeographDashFa35

\u{fa35}: '卑'

CjkCompatibilityIdeographDashFa36

\u{fa36}: '喝'

CjkCompatibilityIdeographDashFa37

\u{fa37}: '嘆'

CjkCompatibilityIdeographDashFa38

\u{fa38}: '器'

CjkCompatibilityIdeographDashFa39

\u{fa39}: '塀'

CjkCompatibilityIdeographDashFa3a

\u{fa3a}: '墨'

CjkCompatibilityIdeographDashFa3b

\u{fa3b}: '層'

CjkCompatibilityIdeographDashFa3c

\u{fa3c}: '屮'

CjkCompatibilityIdeographDashFa3d

\u{fa3d}: '悔'

CjkCompatibilityIdeographDashFa3e

\u{fa3e}: '慨'

CjkCompatibilityIdeographDashFa3f

\u{fa3f}: '憎'

CjkCompatibilityIdeographDashFa40

\u{fa40}: '懲'

CjkCompatibilityIdeographDashFa41

\u{fa41}: '敏'

CjkCompatibilityIdeographDashFa42

\u{fa42}: '既'

CjkCompatibilityIdeographDashFa43

\u{fa43}: '暑'

CjkCompatibilityIdeographDashFa44

\u{fa44}: '梅'

CjkCompatibilityIdeographDashFa45

\u{fa45}: '海'

CjkCompatibilityIdeographDashFa46

\u{fa46}: '渚'

CjkCompatibilityIdeographDashFa47

\u{fa47}: '漢'

CjkCompatibilityIdeographDashFa48

\u{fa48}: '煮'

CjkCompatibilityIdeographDashFa49

\u{fa49}: '爫'

CjkCompatibilityIdeographDashFa4a

\u{fa4a}: '琢'

CjkCompatibilityIdeographDashFa4b

\u{fa4b}: '碑'

CjkCompatibilityIdeographDashFa4c

\u{fa4c}: '社'

CjkCompatibilityIdeographDashFa4d

\u{fa4d}: '祉'

CjkCompatibilityIdeographDashFa4e

\u{fa4e}: '祈'

CjkCompatibilityIdeographDashFa4f

\u{fa4f}: '祐'

CjkCompatibilityIdeographDashFa50

\u{fa50}: '祖'

CjkCompatibilityIdeographDashFa51

\u{fa51}: '祝'

CjkCompatibilityIdeographDashFa52

\u{fa52}: '禍'

CjkCompatibilityIdeographDashFa53

\u{fa53}: '禎'

CjkCompatibilityIdeographDashFa54

\u{fa54}: '穀'

CjkCompatibilityIdeographDashFa55

\u{fa55}: '突'

CjkCompatibilityIdeographDashFa56

\u{fa56}: '節'

CjkCompatibilityIdeographDashFa57

\u{fa57}: '練'

CjkCompatibilityIdeographDashFa58

\u{fa58}: '縉'

CjkCompatibilityIdeographDashFa59

\u{fa59}: '繁'

CjkCompatibilityIdeographDashFa5a

\u{fa5a}: '署'

CjkCompatibilityIdeographDashFa5b

\u{fa5b}: '者'

CjkCompatibilityIdeographDashFa5c

\u{fa5c}: '臭'

CjkCompatibilityIdeographDashFa5d

\u{fa5d}: '艹'

CjkCompatibilityIdeographDashFa5e

\u{fa5e}: '艹'

CjkCompatibilityIdeographDashFa5f

\u{fa5f}: '著'

CjkCompatibilityIdeographDashFa60

\u{fa60}: '褐'

CjkCompatibilityIdeographDashFa61

\u{fa61}: '視'

CjkCompatibilityIdeographDashFa62

\u{fa62}: '謁'

CjkCompatibilityIdeographDashFa63

\u{fa63}: '謹'

CjkCompatibilityIdeographDashFa64

\u{fa64}: '賓'

CjkCompatibilityIdeographDashFa65

\u{fa65}: '贈'

CjkCompatibilityIdeographDashFa66

\u{fa66}: '辶'

CjkCompatibilityIdeographDashFa67

\u{fa67}: '逸'

CjkCompatibilityIdeographDashFa68

\u{fa68}: '難'

CjkCompatibilityIdeographDashFa69

\u{fa69}: '響'

CjkCompatibilityIdeographDashFa6a

\u{fa6a}: '頻'

CjkCompatibilityIdeographDashFa6b

\u{fa6b}: '恵'

CjkCompatibilityIdeographDashFa6c

\u{fa6c}: '𤋮'

CjkCompatibilityIdeographDashFa6d

\u{fa6d}: '舘'

CjkCompatibilityIdeographDashFa70

\u{fa70}: '並'

CjkCompatibilityIdeographDashFa71

\u{fa71}: '况'

CjkCompatibilityIdeographDashFa72

\u{fa72}: '全'

CjkCompatibilityIdeographDashFa73

\u{fa73}: '侀'

CjkCompatibilityIdeographDashFa74

\u{fa74}: '充'

CjkCompatibilityIdeographDashFa75

\u{fa75}: '冀'

CjkCompatibilityIdeographDashFa76

\u{fa76}: '勇'

CjkCompatibilityIdeographDashFa77

\u{fa77}: '勺'

CjkCompatibilityIdeographDashFa78

\u{fa78}: '喝'

CjkCompatibilityIdeographDashFa79

\u{fa79}: '啕'

CjkCompatibilityIdeographDashFa7a

\u{fa7a}: '喙'

CjkCompatibilityIdeographDashFa7b

\u{fa7b}: '嗢'

CjkCompatibilityIdeographDashFa7c

\u{fa7c}: '塚'

CjkCompatibilityIdeographDashFa7d

\u{fa7d}: '墳'

CjkCompatibilityIdeographDashFa7e

\u{fa7e}: '奄'

CjkCompatibilityIdeographDashFa7f

\u{fa7f}: '奔'

CjkCompatibilityIdeographDashFa80

\u{fa80}: '婢'

CjkCompatibilityIdeographDashFa81

\u{fa81}: '嬨'

CjkCompatibilityIdeographDashFa82

\u{fa82}: '廒'

CjkCompatibilityIdeographDashFa83

\u{fa83}: '廙'

CjkCompatibilityIdeographDashFa84

\u{fa84}: '彩'

CjkCompatibilityIdeographDashFa85

\u{fa85}: '徭'

CjkCompatibilityIdeographDashFa86

\u{fa86}: '惘'

CjkCompatibilityIdeographDashFa87

\u{fa87}: '慎'

CjkCompatibilityIdeographDashFa88

\u{fa88}: '愈'

CjkCompatibilityIdeographDashFa89

\u{fa89}: '憎'

CjkCompatibilityIdeographDashFa8a

\u{fa8a}: '慠'

CjkCompatibilityIdeographDashFa8b

\u{fa8b}: '懲'

CjkCompatibilityIdeographDashFa8c

\u{fa8c}: '戴'

CjkCompatibilityIdeographDashFa8d

\u{fa8d}: '揄'

CjkCompatibilityIdeographDashFa8e

\u{fa8e}: '搜'

CjkCompatibilityIdeographDashFa8f

\u{fa8f}: '摒'

CjkCompatibilityIdeographDashFa90

\u{fa90}: '敖'

CjkCompatibilityIdeographDashFa91

\u{fa91}: '晴'

CjkCompatibilityIdeographDashFa92

\u{fa92}: '朗'

CjkCompatibilityIdeographDashFa93

\u{fa93}: '望'

CjkCompatibilityIdeographDashFa94

\u{fa94}: '杖'

CjkCompatibilityIdeographDashFa95

\u{fa95}: '歹'

CjkCompatibilityIdeographDashFa96

\u{fa96}: '殺'

CjkCompatibilityIdeographDashFa97

\u{fa97}: '流'

CjkCompatibilityIdeographDashFa98

\u{fa98}: '滛'

CjkCompatibilityIdeographDashFa99

\u{fa99}: '滋'

CjkCompatibilityIdeographDashFa9a

\u{fa9a}: '漢'

CjkCompatibilityIdeographDashFa9b

\u{fa9b}: '瀞'

CjkCompatibilityIdeographDashFa9c

\u{fa9c}: '煮'

CjkCompatibilityIdeographDashFa9d

\u{fa9d}: '瞧'

CjkCompatibilityIdeographDashFa9e

\u{fa9e}: '爵'

CjkCompatibilityIdeographDashFa9f

\u{fa9f}: '犯'

CjkCompatibilityIdeographDashFaa0

\u{faa0}: '猪'

CjkCompatibilityIdeographDashFaa1

\u{faa1}: '瑱'

CjkCompatibilityIdeographDashFaa2

\u{faa2}: '甆'

CjkCompatibilityIdeographDashFaa3

\u{faa3}: '画'

CjkCompatibilityIdeographDashFaa4

\u{faa4}: '瘝'

CjkCompatibilityIdeographDashFaa5

\u{faa5}: '瘟'

CjkCompatibilityIdeographDashFaa6

\u{faa6}: '益'

CjkCompatibilityIdeographDashFaa7

\u{faa7}: '盛'

CjkCompatibilityIdeographDashFaa8

\u{faa8}: '直'

CjkCompatibilityIdeographDashFaa9

\u{faa9}: '睊'

CjkCompatibilityIdeographDashFaaa

\u{faaa}: '着'

CjkCompatibilityIdeographDashFaab

\u{faab}: '磌'

CjkCompatibilityIdeographDashFaac

\u{faac}: '窱'

CjkCompatibilityIdeographDashFaad

\u{faad}: '節'

CjkCompatibilityIdeographDashFaae

\u{faae}: '类'

CjkCompatibilityIdeographDashFaaf

\u{faaf}: '絛'

CjkCompatibilityIdeographDashFab0

\u{fab0}: '練'

CjkCompatibilityIdeographDashFab1

\u{fab1}: '缾'

CjkCompatibilityIdeographDashFab2

\u{fab2}: '者'

CjkCompatibilityIdeographDashFab3

\u{fab3}: '荒'

CjkCompatibilityIdeographDashFab4

\u{fab4}: '華'

CjkCompatibilityIdeographDashFab5

\u{fab5}: '蝹'

CjkCompatibilityIdeographDashFab6

\u{fab6}: '襁'

CjkCompatibilityIdeographDashFab7

\u{fab7}: '覆'

CjkCompatibilityIdeographDashFab8

\u{fab8}: '視'

CjkCompatibilityIdeographDashFab9

\u{fab9}: '調'

CjkCompatibilityIdeographDashFaba

\u{faba}: '諸'

CjkCompatibilityIdeographDashFabb

\u{fabb}: '請'

CjkCompatibilityIdeographDashFabc

\u{fabc}: '謁'

CjkCompatibilityIdeographDashFabd

\u{fabd}: '諾'

CjkCompatibilityIdeographDashFabe

\u{fabe}: '諭'

CjkCompatibilityIdeographDashFabf

\u{fabf}: '謹'

CjkCompatibilityIdeographDashFac0

\u{fac0}: '變'

CjkCompatibilityIdeographDashFac1

\u{fac1}: '贈'

CjkCompatibilityIdeographDashFac2

\u{fac2}: '輸'

CjkCompatibilityIdeographDashFac3

\u{fac3}: '遲'

CjkCompatibilityIdeographDashFac4

\u{fac4}: '醙'

CjkCompatibilityIdeographDashFac5

\u{fac5}: '鉶'

CjkCompatibilityIdeographDashFac6

\u{fac6}: '陼'

CjkCompatibilityIdeographDashFac7

\u{fac7}: '難'

CjkCompatibilityIdeographDashFac8

\u{fac8}: '靖'

CjkCompatibilityIdeographDashFac9

\u{fac9}: '韛'

CjkCompatibilityIdeographDashFaca

\u{faca}: '響'

CjkCompatibilityIdeographDashFacb

\u{facb}: '頋'

CjkCompatibilityIdeographDashFacc

\u{facc}: '頻'

CjkCompatibilityIdeographDashFacd

\u{facd}: '鬒'

CjkCompatibilityIdeographDashFace

\u{face}: '龜'

CjkCompatibilityIdeographDashFacf

\u{facf}: '𢡊'

CjkCompatibilityIdeographDashFad0

\u{fad0}: '𢡄'

CjkCompatibilityIdeographDashFad1

\u{fad1}: '𣏕'

CjkCompatibilityIdeographDashFad2

\u{fad2}: '㮝'

CjkCompatibilityIdeographDashFad3

\u{fad3}: '䀘'

CjkCompatibilityIdeographDashFad4

\u{fad4}: '䀹'

CjkCompatibilityIdeographDashFad5

\u{fad5}: '𥉉'

CjkCompatibilityIdeographDashFad6

\u{fad6}: '𥳐'

CjkCompatibilityIdeographDashFad7

\u{fad7}: '𧻓'

CjkCompatibilityIdeographDashFad8

\u{fad8}: '齃'

CjkCompatibilityIdeographDashFad9

\u{fad9}: '龎'

Methods

impl CJKCompatibilityIdeographs

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case