Enum unicode_types::cjk_compatibility_ideographs_supplement::CJKCompatibilityIdeographsSupplement

pub enum CJKCompatibilityIdeographsSupplement {
    CjkCompatibilityIdeographDash2f800,
    CjkCompatibilityIdeographDash2f801,
    CjkCompatibilityIdeographDash2f802,
    CjkCompatibilityIdeographDash2f803,
    CjkCompatibilityIdeographDash2f804,
    CjkCompatibilityIdeographDash2f805,
    CjkCompatibilityIdeographDash2f806,
    CjkCompatibilityIdeographDash2f807,
    CjkCompatibilityIdeographDash2f808,
    CjkCompatibilityIdeographDash2f809,
    CjkCompatibilityIdeographDash2f80a,
    CjkCompatibilityIdeographDash2f80b,
    CjkCompatibilityIdeographDash2f80c,
    CjkCompatibilityIdeographDash2f80d,
    CjkCompatibilityIdeographDash2f80e,
    CjkCompatibilityIdeographDash2f80f,
    CjkCompatibilityIdeographDash2f810,
    CjkCompatibilityIdeographDash2f811,
    CjkCompatibilityIdeographDash2f812,
    CjkCompatibilityIdeographDash2f813,
    CjkCompatibilityIdeographDash2f814,
    CjkCompatibilityIdeographDash2f815,
    CjkCompatibilityIdeographDash2f816,
    CjkCompatibilityIdeographDash2f817,
    CjkCompatibilityIdeographDash2f818,
    CjkCompatibilityIdeographDash2f819,
    CjkCompatibilityIdeographDash2f81a,
    CjkCompatibilityIdeographDash2f81b,
    CjkCompatibilityIdeographDash2f81c,
    CjkCompatibilityIdeographDash2f81d,
    CjkCompatibilityIdeographDash2f81e,
    CjkCompatibilityIdeographDash2f81f,
    CjkCompatibilityIdeographDash2f820,
    CjkCompatibilityIdeographDash2f821,
    CjkCompatibilityIdeographDash2f822,
    CjkCompatibilityIdeographDash2f823,
    CjkCompatibilityIdeographDash2f824,
    CjkCompatibilityIdeographDash2f825,
    CjkCompatibilityIdeographDash2f826,
    CjkCompatibilityIdeographDash2f827,
    CjkCompatibilityIdeographDash2f828,
    CjkCompatibilityIdeographDash2f829,
    CjkCompatibilityIdeographDash2f82a,
    CjkCompatibilityIdeographDash2f82b,
    CjkCompatibilityIdeographDash2f82c,
    CjkCompatibilityIdeographDash2f82d,
    CjkCompatibilityIdeographDash2f82e,
    CjkCompatibilityIdeographDash2f82f,
    CjkCompatibilityIdeographDash2f830,
    CjkCompatibilityIdeographDash2f831,
    CjkCompatibilityIdeographDash2f832,
    CjkCompatibilityIdeographDash2f833,
    CjkCompatibilityIdeographDash2f834,
    CjkCompatibilityIdeographDash2f835,
    CjkCompatibilityIdeographDash2f836,
    CjkCompatibilityIdeographDash2f837,
    CjkCompatibilityIdeographDash2f838,
    CjkCompatibilityIdeographDash2f839,
    CjkCompatibilityIdeographDash2f83a,
    CjkCompatibilityIdeographDash2f83b,
    CjkCompatibilityIdeographDash2f83c,
    CjkCompatibilityIdeographDash2f83d,
    CjkCompatibilityIdeographDash2f83e,
    CjkCompatibilityIdeographDash2f83f,
    CjkCompatibilityIdeographDash2f840,
    CjkCompatibilityIdeographDash2f841,
    CjkCompatibilityIdeographDash2f842,
    CjkCompatibilityIdeographDash2f843,
    CjkCompatibilityIdeographDash2f844,
    CjkCompatibilityIdeographDash2f845,
    CjkCompatibilityIdeographDash2f846,
    CjkCompatibilityIdeographDash2f847,
    CjkCompatibilityIdeographDash2f848,
    CjkCompatibilityIdeographDash2f849,
    CjkCompatibilityIdeographDash2f84a,
    CjkCompatibilityIdeographDash2f84b,
    CjkCompatibilityIdeographDash2f84c,
    CjkCompatibilityIdeographDash2f84d,
    CjkCompatibilityIdeographDash2f84e,
    CjkCompatibilityIdeographDash2f84f,
    CjkCompatibilityIdeographDash2f850,
    CjkCompatibilityIdeographDash2f851,
    CjkCompatibilityIdeographDash2f852,
    CjkCompatibilityIdeographDash2f853,
    CjkCompatibilityIdeographDash2f854,
    CjkCompatibilityIdeographDash2f855,
    CjkCompatibilityIdeographDash2f856,
    CjkCompatibilityIdeographDash2f857,
    CjkCompatibilityIdeographDash2f858,
    CjkCompatibilityIdeographDash2f859,
    CjkCompatibilityIdeographDash2f85a,
    CjkCompatibilityIdeographDash2f85b,
    CjkCompatibilityIdeographDash2f85c,
    CjkCompatibilityIdeographDash2f85d,
    CjkCompatibilityIdeographDash2f85e,
    CjkCompatibilityIdeographDash2f85f,
    CjkCompatibilityIdeographDash2f860,
    CjkCompatibilityIdeographDash2f861,
    CjkCompatibilityIdeographDash2f862,
    CjkCompatibilityIdeographDash2f863,
    CjkCompatibilityIdeographDash2f864,
    CjkCompatibilityIdeographDash2f865,
    CjkCompatibilityIdeographDash2f866,
    CjkCompatibilityIdeographDash2f867,
    CjkCompatibilityIdeographDash2f868,
    CjkCompatibilityIdeographDash2f869,
    CjkCompatibilityIdeographDash2f86a,
    CjkCompatibilityIdeographDash2f86b,
    CjkCompatibilityIdeographDash2f86c,
    CjkCompatibilityIdeographDash2f86d,
    CjkCompatibilityIdeographDash2f86e,
    CjkCompatibilityIdeographDash2f86f,
    CjkCompatibilityIdeographDash2f870,
    CjkCompatibilityIdeographDash2f871,
    CjkCompatibilityIdeographDash2f872,
    CjkCompatibilityIdeographDash2f873,
    CjkCompatibilityIdeographDash2f874,
    CjkCompatibilityIdeographDash2f875,
    CjkCompatibilityIdeographDash2f876,
    CjkCompatibilityIdeographDash2f877,
    CjkCompatibilityIdeographDash2f878,
    CjkCompatibilityIdeographDash2f879,
    CjkCompatibilityIdeographDash2f87a,
    CjkCompatibilityIdeographDash2f87b,
    CjkCompatibilityIdeographDash2f87c,
    CjkCompatibilityIdeographDash2f87d,
    CjkCompatibilityIdeographDash2f87e,
    CjkCompatibilityIdeographDash2f87f,
    CjkCompatibilityIdeographDash2f880,
    CjkCompatibilityIdeographDash2f881,
    CjkCompatibilityIdeographDash2f882,
    CjkCompatibilityIdeographDash2f883,
    CjkCompatibilityIdeographDash2f884,
    CjkCompatibilityIdeographDash2f885,
    CjkCompatibilityIdeographDash2f886,
    CjkCompatibilityIdeographDash2f887,
    CjkCompatibilityIdeographDash2f888,
    CjkCompatibilityIdeographDash2f889,
    CjkCompatibilityIdeographDash2f88a,
    CjkCompatibilityIdeographDash2f88b,
    CjkCompatibilityIdeographDash2f88c,
    CjkCompatibilityIdeographDash2f88d,
    CjkCompatibilityIdeographDash2f88e,
    CjkCompatibilityIdeographDash2f88f,
    CjkCompatibilityIdeographDash2f890,
    CjkCompatibilityIdeographDash2f891,
    CjkCompatibilityIdeographDash2f892,
    CjkCompatibilityIdeographDash2f893,
    CjkCompatibilityIdeographDash2f894,
    CjkCompatibilityIdeographDash2f895,
    CjkCompatibilityIdeographDash2f896,
    CjkCompatibilityIdeographDash2f897,
    CjkCompatibilityIdeographDash2f898,
    CjkCompatibilityIdeographDash2f899,
    CjkCompatibilityIdeographDash2f89a,
    CjkCompatibilityIdeographDash2f89b,
    CjkCompatibilityIdeographDash2f89c,
    CjkCompatibilityIdeographDash2f89d,
    CjkCompatibilityIdeographDash2f89e,
    CjkCompatibilityIdeographDash2f89f,
    CjkCompatibilityIdeographDash2f8a0,
    CjkCompatibilityIdeographDash2f8a1,
    CjkCompatibilityIdeographDash2f8a2,
    CjkCompatibilityIdeographDash2f8a3,
    CjkCompatibilityIdeographDash2f8a4,
    CjkCompatibilityIdeographDash2f8a5,
    CjkCompatibilityIdeographDash2f8a6,
    CjkCompatibilityIdeographDash2f8a7,
    CjkCompatibilityIdeographDash2f8a8,
    CjkCompatibilityIdeographDash2f8a9,
    CjkCompatibilityIdeographDash2f8aa,
    CjkCompatibilityIdeographDash2f8ab,
    CjkCompatibilityIdeographDash2f8ac,
    CjkCompatibilityIdeographDash2f8ad,
    CjkCompatibilityIdeographDash2f8ae,
    CjkCompatibilityIdeographDash2f8af,
    CjkCompatibilityIdeographDash2f8b0,
    CjkCompatibilityIdeographDash2f8b1,
    CjkCompatibilityIdeographDash2f8b2,
    CjkCompatibilityIdeographDash2f8b3,
    CjkCompatibilityIdeographDash2f8b4,
    CjkCompatibilityIdeographDash2f8b5,
    CjkCompatibilityIdeographDash2f8b6,
    CjkCompatibilityIdeographDash2f8b7,
    CjkCompatibilityIdeographDash2f8b8,
    CjkCompatibilityIdeographDash2f8b9,
    CjkCompatibilityIdeographDash2f8ba,
    CjkCompatibilityIdeographDash2f8bb,
    CjkCompatibilityIdeographDash2f8bc,
    CjkCompatibilityIdeographDash2f8bd,
    CjkCompatibilityIdeographDash2f8be,
    CjkCompatibilityIdeographDash2f8bf,
    CjkCompatibilityIdeographDash2f8c0,
    CjkCompatibilityIdeographDash2f8c1,
    CjkCompatibilityIdeographDash2f8c2,
    CjkCompatibilityIdeographDash2f8c3,
    CjkCompatibilityIdeographDash2f8c4,
    CjkCompatibilityIdeographDash2f8c5,
    CjkCompatibilityIdeographDash2f8c6,
    CjkCompatibilityIdeographDash2f8c7,
    CjkCompatibilityIdeographDash2f8c8,
    CjkCompatibilityIdeographDash2f8c9,
    CjkCompatibilityIdeographDash2f8ca,
    CjkCompatibilityIdeographDash2f8cb,
    CjkCompatibilityIdeographDash2f8cc,
    CjkCompatibilityIdeographDash2f8cd,
    CjkCompatibilityIdeographDash2f8ce,
    CjkCompatibilityIdeographDash2f8cf,
    CjkCompatibilityIdeographDash2f8d0,
    CjkCompatibilityIdeographDash2f8d1,
    CjkCompatibilityIdeographDash2f8d2,
    CjkCompatibilityIdeographDash2f8d3,
    CjkCompatibilityIdeographDash2f8d4,
    CjkCompatibilityIdeographDash2f8d5,
    CjkCompatibilityIdeographDash2f8d6,
    CjkCompatibilityIdeographDash2f8d7,
    CjkCompatibilityIdeographDash2f8d8,
    CjkCompatibilityIdeographDash2f8d9,
    CjkCompatibilityIdeographDash2f8da,
    CjkCompatibilityIdeographDash2f8db,
    CjkCompatibilityIdeographDash2f8dc,
    CjkCompatibilityIdeographDash2f8dd,
    CjkCompatibilityIdeographDash2f8de,
    CjkCompatibilityIdeographDash2f8df,
    CjkCompatibilityIdeographDash2f8e0,
    CjkCompatibilityIdeographDash2f8e1,
    CjkCompatibilityIdeographDash2f8e2,
    CjkCompatibilityIdeographDash2f8e3,
    CjkCompatibilityIdeographDash2f8e4,
    CjkCompatibilityIdeographDash2f8e5,
    CjkCompatibilityIdeographDash2f8e6,
    CjkCompatibilityIdeographDash2f8e7,
    CjkCompatibilityIdeographDash2f8e8,
    CjkCompatibilityIdeographDash2f8e9,
    CjkCompatibilityIdeographDash2f8ea,
    CjkCompatibilityIdeographDash2f8eb,
    CjkCompatibilityIdeographDash2f8ec,
    CjkCompatibilityIdeographDash2f8ed,
    CjkCompatibilityIdeographDash2f8ee,
    CjkCompatibilityIdeographDash2f8ef,
    CjkCompatibilityIdeographDash2f8f0,
    CjkCompatibilityIdeographDash2f8f1,
    CjkCompatibilityIdeographDash2f8f2,
    CjkCompatibilityIdeographDash2f8f3,
    CjkCompatibilityIdeographDash2f8f4,
    CjkCompatibilityIdeographDash2f8f5,
    CjkCompatibilityIdeographDash2f8f6,
    CjkCompatibilityIdeographDash2f8f7,
    CjkCompatibilityIdeographDash2f8f8,
    CjkCompatibilityIdeographDash2f8f9,
    CjkCompatibilityIdeographDash2f8fa,
    CjkCompatibilityIdeographDash2f8fb,
    CjkCompatibilityIdeographDash2f8fc,
    CjkCompatibilityIdeographDash2f8fd,
    CjkCompatibilityIdeographDash2f8fe,
    CjkCompatibilityIdeographDash2f8ff,
    CjkCompatibilityIdeographDash2f900,
    CjkCompatibilityIdeographDash2f901,
    CjkCompatibilityIdeographDash2f902,
    CjkCompatibilityIdeographDash2f903,
    CjkCompatibilityIdeographDash2f904,
    CjkCompatibilityIdeographDash2f905,
    CjkCompatibilityIdeographDash2f906,
    CjkCompatibilityIdeographDash2f907,
    CjkCompatibilityIdeographDash2f908,
    CjkCompatibilityIdeographDash2f909,
    CjkCompatibilityIdeographDash2f90a,
    CjkCompatibilityIdeographDash2f90b,
    CjkCompatibilityIdeographDash2f90c,
    CjkCompatibilityIdeographDash2f90d,
    CjkCompatibilityIdeographDash2f90e,
    CjkCompatibilityIdeographDash2f90f,
    CjkCompatibilityIdeographDash2f910,
    CjkCompatibilityIdeographDash2f911,
    CjkCompatibilityIdeographDash2f912,
    CjkCompatibilityIdeographDash2f913,
    CjkCompatibilityIdeographDash2f914,
    CjkCompatibilityIdeographDash2f915,
    CjkCompatibilityIdeographDash2f916,
    CjkCompatibilityIdeographDash2f917,
    CjkCompatibilityIdeographDash2f918,
    CjkCompatibilityIdeographDash2f919,
    CjkCompatibilityIdeographDash2f91a,
    CjkCompatibilityIdeographDash2f91b,
    CjkCompatibilityIdeographDash2f91c,
    CjkCompatibilityIdeographDash2f91d,
    CjkCompatibilityIdeographDash2f91e,
    CjkCompatibilityIdeographDash2f91f,
    CjkCompatibilityIdeographDash2f920,
    CjkCompatibilityIdeographDash2f921,
    CjkCompatibilityIdeographDash2f922,
    CjkCompatibilityIdeographDash2f923,
    CjkCompatibilityIdeographDash2f924,
    CjkCompatibilityIdeographDash2f925,
    CjkCompatibilityIdeographDash2f926,
    CjkCompatibilityIdeographDash2f927,
    CjkCompatibilityIdeographDash2f928,
    CjkCompatibilityIdeographDash2f929,
    CjkCompatibilityIdeographDash2f92a,
    CjkCompatibilityIdeographDash2f92b,
    CjkCompatibilityIdeographDash2f92c,
    CjkCompatibilityIdeographDash2f92d,
    CjkCompatibilityIdeographDash2f92e,
    CjkCompatibilityIdeographDash2f92f,
    CjkCompatibilityIdeographDash2f930,
    CjkCompatibilityIdeographDash2f931,
    CjkCompatibilityIdeographDash2f932,
    CjkCompatibilityIdeographDash2f933,
    CjkCompatibilityIdeographDash2f934,
    CjkCompatibilityIdeographDash2f935,
    CjkCompatibilityIdeographDash2f936,
    CjkCompatibilityIdeographDash2f937,
    CjkCompatibilityIdeographDash2f938,
    CjkCompatibilityIdeographDash2f939,
    CjkCompatibilityIdeographDash2f93a,
    CjkCompatibilityIdeographDash2f93b,
    CjkCompatibilityIdeographDash2f93c,
    CjkCompatibilityIdeographDash2f93d,
    CjkCompatibilityIdeographDash2f93e,
    CjkCompatibilityIdeographDash2f93f,
    CjkCompatibilityIdeographDash2f940,
    CjkCompatibilityIdeographDash2f941,
    CjkCompatibilityIdeographDash2f942,
    CjkCompatibilityIdeographDash2f943,
    CjkCompatibilityIdeographDash2f944,
    CjkCompatibilityIdeographDash2f945,
    CjkCompatibilityIdeographDash2f946,
    CjkCompatibilityIdeographDash2f947,
    CjkCompatibilityIdeographDash2f948,
    CjkCompatibilityIdeographDash2f949,
    CjkCompatibilityIdeographDash2f94a,
    CjkCompatibilityIdeographDash2f94b,
    CjkCompatibilityIdeographDash2f94c,
    CjkCompatibilityIdeographDash2f94d,
    CjkCompatibilityIdeographDash2f94e,
    CjkCompatibilityIdeographDash2f94f,
    CjkCompatibilityIdeographDash2f950,
    CjkCompatibilityIdeographDash2f951,
    CjkCompatibilityIdeographDash2f952,
    CjkCompatibilityIdeographDash2f953,
    CjkCompatibilityIdeographDash2f954,
    CjkCompatibilityIdeographDash2f955,
    CjkCompatibilityIdeographDash2f956,
    CjkCompatibilityIdeographDash2f957,
    CjkCompatibilityIdeographDash2f958,
    CjkCompatibilityIdeographDash2f959,
    CjkCompatibilityIdeographDash2f95a,
    CjkCompatibilityIdeographDash2f95b,
    CjkCompatibilityIdeographDash2f95c,
    CjkCompatibilityIdeographDash2f95d,
    CjkCompatibilityIdeographDash2f95e,
    CjkCompatibilityIdeographDash2f95f,
    CjkCompatibilityIdeographDash2f960,
    CjkCompatibilityIdeographDash2f961,
    CjkCompatibilityIdeographDash2f962,
    CjkCompatibilityIdeographDash2f963,
    CjkCompatibilityIdeographDash2f964,
    CjkCompatibilityIdeographDash2f965,
    CjkCompatibilityIdeographDash2f966,
    CjkCompatibilityIdeographDash2f967,
    CjkCompatibilityIdeographDash2f968,
    CjkCompatibilityIdeographDash2f969,
    CjkCompatibilityIdeographDash2f96a,
    CjkCompatibilityIdeographDash2f96b,
    CjkCompatibilityIdeographDash2f96c,
    CjkCompatibilityIdeographDash2f96d,
    CjkCompatibilityIdeographDash2f96e,
    CjkCompatibilityIdeographDash2f96f,
    CjkCompatibilityIdeographDash2f970,
    CjkCompatibilityIdeographDash2f971,
    CjkCompatibilityIdeographDash2f972,
    CjkCompatibilityIdeographDash2f973,
    CjkCompatibilityIdeographDash2f974,
    CjkCompatibilityIdeographDash2f975,
    CjkCompatibilityIdeographDash2f976,
    CjkCompatibilityIdeographDash2f977,
    CjkCompatibilityIdeographDash2f978,
    CjkCompatibilityIdeographDash2f979,
    CjkCompatibilityIdeographDash2f97a,
    CjkCompatibilityIdeographDash2f97b,
    CjkCompatibilityIdeographDash2f97c,
    CjkCompatibilityIdeographDash2f97d,
    CjkCompatibilityIdeographDash2f97e,
    CjkCompatibilityIdeographDash2f97f,
    CjkCompatibilityIdeographDash2f980,
    CjkCompatibilityIdeographDash2f981,
    CjkCompatibilityIdeographDash2f982,
    CjkCompatibilityIdeographDash2f983,
    CjkCompatibilityIdeographDash2f984,
    CjkCompatibilityIdeographDash2f985,
    CjkCompatibilityIdeographDash2f986,
    CjkCompatibilityIdeographDash2f987,
    CjkCompatibilityIdeographDash2f988,
    CjkCompatibilityIdeographDash2f989,
    CjkCompatibilityIdeographDash2f98a,
    CjkCompatibilityIdeographDash2f98b,
    CjkCompatibilityIdeographDash2f98c,
    CjkCompatibilityIdeographDash2f98d,
    CjkCompatibilityIdeographDash2f98e,
    CjkCompatibilityIdeographDash2f98f,
    CjkCompatibilityIdeographDash2f990,
    CjkCompatibilityIdeographDash2f991,
    CjkCompatibilityIdeographDash2f992,
    CjkCompatibilityIdeographDash2f993,
    CjkCompatibilityIdeographDash2f994,
    CjkCompatibilityIdeographDash2f995,
    CjkCompatibilityIdeographDash2f996,
    CjkCompatibilityIdeographDash2f997,
    CjkCompatibilityIdeographDash2f998,
    CjkCompatibilityIdeographDash2f999,
    CjkCompatibilityIdeographDash2f99a,
    CjkCompatibilityIdeographDash2f99b,
    CjkCompatibilityIdeographDash2f99c,
    CjkCompatibilityIdeographDash2f99d,
    CjkCompatibilityIdeographDash2f99e,
    CjkCompatibilityIdeographDash2f99f,
    CjkCompatibilityIdeographDash2f9a0,
    CjkCompatibilityIdeographDash2f9a1,
    CjkCompatibilityIdeographDash2f9a2,
    CjkCompatibilityIdeographDash2f9a3,
    CjkCompatibilityIdeographDash2f9a4,
    CjkCompatibilityIdeographDash2f9a5,
    CjkCompatibilityIdeographDash2f9a6,
    CjkCompatibilityIdeographDash2f9a7,
    CjkCompatibilityIdeographDash2f9a8,
    CjkCompatibilityIdeographDash2f9a9,
    CjkCompatibilityIdeographDash2f9aa,
    CjkCompatibilityIdeographDash2f9ab,
    CjkCompatibilityIdeographDash2f9ac,
    CjkCompatibilityIdeographDash2f9ad,
    CjkCompatibilityIdeographDash2f9ae,
    CjkCompatibilityIdeographDash2f9af,
    CjkCompatibilityIdeographDash2f9b0,
    CjkCompatibilityIdeographDash2f9b1,
    CjkCompatibilityIdeographDash2f9b2,
    CjkCompatibilityIdeographDash2f9b3,
    CjkCompatibilityIdeographDash2f9b4,
    CjkCompatibilityIdeographDash2f9b5,
    CjkCompatibilityIdeographDash2f9b6,
    CjkCompatibilityIdeographDash2f9b7,
    CjkCompatibilityIdeographDash2f9b8,
    CjkCompatibilityIdeographDash2f9b9,
    CjkCompatibilityIdeographDash2f9ba,
    CjkCompatibilityIdeographDash2f9bb,
    CjkCompatibilityIdeographDash2f9bc,
    CjkCompatibilityIdeographDash2f9bd,
    CjkCompatibilityIdeographDash2f9be,
    CjkCompatibilityIdeographDash2f9bf,
    CjkCompatibilityIdeographDash2f9c0,
    CjkCompatibilityIdeographDash2f9c1,
    CjkCompatibilityIdeographDash2f9c2,
    CjkCompatibilityIdeographDash2f9c3,
    CjkCompatibilityIdeographDash2f9c4,
    CjkCompatibilityIdeographDash2f9c5,
    CjkCompatibilityIdeographDash2f9c6,
    CjkCompatibilityIdeographDash2f9c7,
    CjkCompatibilityIdeographDash2f9c8,
    CjkCompatibilityIdeographDash2f9c9,
    CjkCompatibilityIdeographDash2f9ca,
    CjkCompatibilityIdeographDash2f9cb,
    CjkCompatibilityIdeographDash2f9cc,
    CjkCompatibilityIdeographDash2f9cd,
    CjkCompatibilityIdeographDash2f9ce,
    CjkCompatibilityIdeographDash2f9cf,
    CjkCompatibilityIdeographDash2f9d0,
    CjkCompatibilityIdeographDash2f9d1,
    CjkCompatibilityIdeographDash2f9d2,
    CjkCompatibilityIdeographDash2f9d3,
    CjkCompatibilityIdeographDash2f9d4,
    CjkCompatibilityIdeographDash2f9d5,
    CjkCompatibilityIdeographDash2f9d6,
    CjkCompatibilityIdeographDash2f9d7,
    CjkCompatibilityIdeographDash2f9d8,
    CjkCompatibilityIdeographDash2f9d9,
    CjkCompatibilityIdeographDash2f9da,
    CjkCompatibilityIdeographDash2f9db,
    CjkCompatibilityIdeographDash2f9dc,
    CjkCompatibilityIdeographDash2f9dd,
    CjkCompatibilityIdeographDash2f9de,
    CjkCompatibilityIdeographDash2f9df,
    CjkCompatibilityIdeographDash2f9e0,
    CjkCompatibilityIdeographDash2f9e1,
    CjkCompatibilityIdeographDash2f9e2,
    CjkCompatibilityIdeographDash2f9e3,
    CjkCompatibilityIdeographDash2f9e4,
    CjkCompatibilityIdeographDash2f9e5,
    CjkCompatibilityIdeographDash2f9e6,
    CjkCompatibilityIdeographDash2f9e7,
    CjkCompatibilityIdeographDash2f9e8,
    CjkCompatibilityIdeographDash2f9e9,
    CjkCompatibilityIdeographDash2f9ea,
    CjkCompatibilityIdeographDash2f9eb,
    CjkCompatibilityIdeographDash2f9ec,
    CjkCompatibilityIdeographDash2f9ed,
    CjkCompatibilityIdeographDash2f9ee,
    CjkCompatibilityIdeographDash2f9ef,
    CjkCompatibilityIdeographDash2f9f0,
    CjkCompatibilityIdeographDash2f9f1,
    CjkCompatibilityIdeographDash2f9f2,
    CjkCompatibilityIdeographDash2f9f3,
    CjkCompatibilityIdeographDash2f9f4,
    CjkCompatibilityIdeographDash2f9f5,
    CjkCompatibilityIdeographDash2f9f6,
    CjkCompatibilityIdeographDash2f9f7,
    CjkCompatibilityIdeographDash2f9f8,
    CjkCompatibilityIdeographDash2f9f9,
    CjkCompatibilityIdeographDash2f9fa,
    CjkCompatibilityIdeographDash2f9fb,
    CjkCompatibilityIdeographDash2f9fc,
    CjkCompatibilityIdeographDash2f9fd,
    CjkCompatibilityIdeographDash2f9fe,
    CjkCompatibilityIdeographDash2f9ff,
    CjkCompatibilityIdeographDash2fa00,
    CjkCompatibilityIdeographDash2fa01,
    CjkCompatibilityIdeographDash2fa02,
    CjkCompatibilityIdeographDash2fa03,
    CjkCompatibilityIdeographDash2fa04,
    CjkCompatibilityIdeographDash2fa05,
    CjkCompatibilityIdeographDash2fa06,
    CjkCompatibilityIdeographDash2fa07,
    CjkCompatibilityIdeographDash2fa08,
    CjkCompatibilityIdeographDash2fa09,
    CjkCompatibilityIdeographDash2fa0a,
    CjkCompatibilityIdeographDash2fa0b,
    CjkCompatibilityIdeographDash2fa0c,
    CjkCompatibilityIdeographDash2fa0d,
    CjkCompatibilityIdeographDash2fa0e,
    CjkCompatibilityIdeographDash2fa0f,
    CjkCompatibilityIdeographDash2fa10,
    CjkCompatibilityIdeographDash2fa11,
    CjkCompatibilityIdeographDash2fa12,
    CjkCompatibilityIdeographDash2fa13,
    CjkCompatibilityIdeographDash2fa14,
    CjkCompatibilityIdeographDash2fa15,
    CjkCompatibilityIdeographDash2fa16,
    CjkCompatibilityIdeographDash2fa17,
    CjkCompatibilityIdeographDash2fa18,
    CjkCompatibilityIdeographDash2fa19,
    CjkCompatibilityIdeographDash2fa1a,
    CjkCompatibilityIdeographDash2fa1b,
    CjkCompatibilityIdeographDash2fa1c,
    CjkCompatibilityIdeographDash2fa1d,
}

An enum to represent all characters in the CJKCompatibilityIdeographsSupplement block.

Variants
CjkCompatibilityIdeographDash2f800

\u{2f800}: ''

CjkCompatibilityIdeographDash2f801

\u{2f801}: ''

CjkCompatibilityIdeographDash2f802

\u{2f802}: ''

CjkCompatibilityIdeographDash2f803

\u{2f803}: ''

CjkCompatibilityIdeographDash2f804

\u{2f804}: ''

CjkCompatibilityIdeographDash2f805

\u{2f805}: ''

CjkCompatibilityIdeographDash2f806

\u{2f806}: ''

CjkCompatibilityIdeographDash2f807

\u{2f807}: ''

CjkCompatibilityIdeographDash2f808

\u{2f808}: ''

CjkCompatibilityIdeographDash2f809

\u{2f809}: ''

CjkCompatibilityIdeographDash2f80a

\u{2f80a}: ''

CjkCompatibilityIdeographDash2f80b

\u{2f80b}: ''

CjkCompatibilityIdeographDash2f80c

\u{2f80c}: ''

CjkCompatibilityIdeographDash2f80d

\u{2f80d}: ''

CjkCompatibilityIdeographDash2f80e

\u{2f80e}: ''

CjkCompatibilityIdeographDash2f80f

\u{2f80f}: ''

CjkCompatibilityIdeographDash2f810

\u{2f810}: ''

CjkCompatibilityIdeographDash2f811

\u{2f811}: ''

CjkCompatibilityIdeographDash2f812

\u{2f812}: ''

CjkCompatibilityIdeographDash2f813

\u{2f813}: ''

CjkCompatibilityIdeographDash2f814

\u{2f814}: ''

CjkCompatibilityIdeographDash2f815

\u{2f815}: ''

CjkCompatibilityIdeographDash2f816

\u{2f816}: ''

CjkCompatibilityIdeographDash2f817

\u{2f817}: ''

CjkCompatibilityIdeographDash2f818

\u{2f818}: ''

CjkCompatibilityIdeographDash2f819

\u{2f819}: ''

CjkCompatibilityIdeographDash2f81a

\u{2f81a}: ''

CjkCompatibilityIdeographDash2f81b

\u{2f81b}: ''

CjkCompatibilityIdeographDash2f81c

\u{2f81c}: ''

CjkCompatibilityIdeographDash2f81d

\u{2f81d}: ''

CjkCompatibilityIdeographDash2f81e

\u{2f81e}: ''

CjkCompatibilityIdeographDash2f81f

\u{2f81f}: ''

CjkCompatibilityIdeographDash2f820

\u{2f820}: ''

CjkCompatibilityIdeographDash2f821

\u{2f821}: ''

CjkCompatibilityIdeographDash2f822

\u{2f822}: ''

CjkCompatibilityIdeographDash2f823

\u{2f823}: ''

CjkCompatibilityIdeographDash2f824

\u{2f824}: ''

CjkCompatibilityIdeographDash2f825

\u{2f825}: ''

CjkCompatibilityIdeographDash2f826

\u{2f826}: ''

CjkCompatibilityIdeographDash2f827

\u{2f827}: ''

CjkCompatibilityIdeographDash2f828

\u{2f828}: ''

CjkCompatibilityIdeographDash2f829

\u{2f829}: ''

CjkCompatibilityIdeographDash2f82a

\u{2f82a}: ''

CjkCompatibilityIdeographDash2f82b

\u{2f82b}: ''

CjkCompatibilityIdeographDash2f82c

\u{2f82c}: ''

CjkCompatibilityIdeographDash2f82d

\u{2f82d}: ''

CjkCompatibilityIdeographDash2f82e

\u{2f82e}: ''

CjkCompatibilityIdeographDash2f82f

\u{2f82f}: ''

CjkCompatibilityIdeographDash2f830

\u{2f830}: ''

CjkCompatibilityIdeographDash2f831

\u{2f831}: ''

CjkCompatibilityIdeographDash2f832

\u{2f832}: ''

CjkCompatibilityIdeographDash2f833

\u{2f833}: ''

CjkCompatibilityIdeographDash2f834

\u{2f834}: ''

CjkCompatibilityIdeographDash2f835

\u{2f835}: ''

CjkCompatibilityIdeographDash2f836

\u{2f836}: ''

CjkCompatibilityIdeographDash2f837

\u{2f837}: ''

CjkCompatibilityIdeographDash2f838

\u{2f838}: ''

CjkCompatibilityIdeographDash2f839

\u{2f839}: ''

CjkCompatibilityIdeographDash2f83a

\u{2f83a}: ''

CjkCompatibilityIdeographDash2f83b

\u{2f83b}: ''

CjkCompatibilityIdeographDash2f83c

\u{2f83c}: ''

CjkCompatibilityIdeographDash2f83d

\u{2f83d}: ''

CjkCompatibilityIdeographDash2f83e

\u{2f83e}: ''

CjkCompatibilityIdeographDash2f83f

\u{2f83f}: ''

CjkCompatibilityIdeographDash2f840

\u{2f840}: ''

CjkCompatibilityIdeographDash2f841

\u{2f841}: ''

CjkCompatibilityIdeographDash2f842

\u{2f842}: ''

CjkCompatibilityIdeographDash2f843

\u{2f843}: ''

CjkCompatibilityIdeographDash2f844

\u{2f844}: ''

CjkCompatibilityIdeographDash2f845

\u{2f845}: ''

CjkCompatibilityIdeographDash2f846

\u{2f846}: ''

CjkCompatibilityIdeographDash2f847

\u{2f847}: ''

CjkCompatibilityIdeographDash2f848

\u{2f848}: ''

CjkCompatibilityIdeographDash2f849

\u{2f849}: ''

CjkCompatibilityIdeographDash2f84a

\u{2f84a}: ''

CjkCompatibilityIdeographDash2f84b

\u{2f84b}: ''

CjkCompatibilityIdeographDash2f84c

\u{2f84c}: ''

CjkCompatibilityIdeographDash2f84d

\u{2f84d}: ''

CjkCompatibilityIdeographDash2f84e

\u{2f84e}: ''

CjkCompatibilityIdeographDash2f84f

\u{2f84f}: ''

CjkCompatibilityIdeographDash2f850

\u{2f850}: ''

CjkCompatibilityIdeographDash2f851

\u{2f851}: ''

CjkCompatibilityIdeographDash2f852

\u{2f852}: ''

CjkCompatibilityIdeographDash2f853

\u{2f853}: ''

CjkCompatibilityIdeographDash2f854

\u{2f854}: ''

CjkCompatibilityIdeographDash2f855

\u{2f855}: ''

CjkCompatibilityIdeographDash2f856

\u{2f856}: ''

CjkCompatibilityIdeographDash2f857

\u{2f857}: ''

CjkCompatibilityIdeographDash2f858

\u{2f858}: ''

CjkCompatibilityIdeographDash2f859

\u{2f859}: ''

CjkCompatibilityIdeographDash2f85a

\u{2f85a}: ''

CjkCompatibilityIdeographDash2f85b

\u{2f85b}: ''

CjkCompatibilityIdeographDash2f85c

\u{2f85c}: ''

CjkCompatibilityIdeographDash2f85d

\u{2f85d}: ''

CjkCompatibilityIdeographDash2f85e

\u{2f85e}: ''

CjkCompatibilityIdeographDash2f85f

\u{2f85f}: ''

CjkCompatibilityIdeographDash2f860

\u{2f860}: ''

CjkCompatibilityIdeographDash2f861

\u{2f861}: ''

CjkCompatibilityIdeographDash2f862

\u{2f862}: ''

CjkCompatibilityIdeographDash2f863

\u{2f863}: ''

CjkCompatibilityIdeographDash2f864

\u{2f864}: ''

CjkCompatibilityIdeographDash2f865

\u{2f865}: ''

CjkCompatibilityIdeographDash2f866

\u{2f866}: ''

CjkCompatibilityIdeographDash2f867

\u{2f867}: ''

CjkCompatibilityIdeographDash2f868

\u{2f868}: ''

CjkCompatibilityIdeographDash2f869

\u{2f869}: ''

CjkCompatibilityIdeographDash2f86a

\u{2f86a}: ''

CjkCompatibilityIdeographDash2f86b

\u{2f86b}: ''

CjkCompatibilityIdeographDash2f86c

\u{2f86c}: ''

CjkCompatibilityIdeographDash2f86d

\u{2f86d}: ''

CjkCompatibilityIdeographDash2f86e

\u{2f86e}: ''

CjkCompatibilityIdeographDash2f86f

\u{2f86f}: ''

CjkCompatibilityIdeographDash2f870

\u{2f870}: ''

CjkCompatibilityIdeographDash2f871

\u{2f871}: ''

CjkCompatibilityIdeographDash2f872

\u{2f872}: ''

CjkCompatibilityIdeographDash2f873

\u{2f873}: ''

CjkCompatibilityIdeographDash2f874

\u{2f874}: ''

CjkCompatibilityIdeographDash2f875

\u{2f875}: ''

CjkCompatibilityIdeographDash2f876

\u{2f876}: ''

CjkCompatibilityIdeographDash2f877

\u{2f877}: ''

CjkCompatibilityIdeographDash2f878

\u{2f878}: ''

CjkCompatibilityIdeographDash2f879

\u{2f879}: ''

CjkCompatibilityIdeographDash2f87a

\u{2f87a}: ''

CjkCompatibilityIdeographDash2f87b

\u{2f87b}: ''

CjkCompatibilityIdeographDash2f87c

\u{2f87c}: ''

CjkCompatibilityIdeographDash2f87d

\u{2f87d}: ''

CjkCompatibilityIdeographDash2f87e

\u{2f87e}: ''

CjkCompatibilityIdeographDash2f87f

\u{2f87f}: ''

CjkCompatibilityIdeographDash2f880

\u{2f880}: ''

CjkCompatibilityIdeographDash2f881

\u{2f881}: ''

CjkCompatibilityIdeographDash2f882

\u{2f882}: ''

CjkCompatibilityIdeographDash2f883

\u{2f883}: ''

CjkCompatibilityIdeographDash2f884

\u{2f884}: ''

CjkCompatibilityIdeographDash2f885

\u{2f885}: ''

CjkCompatibilityIdeographDash2f886

\u{2f886}: ''

CjkCompatibilityIdeographDash2f887

\u{2f887}: ''

CjkCompatibilityIdeographDash2f888

\u{2f888}: ''

CjkCompatibilityIdeographDash2f889

\u{2f889}: ''

CjkCompatibilityIdeographDash2f88a

\u{2f88a}: ''

CjkCompatibilityIdeographDash2f88b

\u{2f88b}: ''

CjkCompatibilityIdeographDash2f88c

\u{2f88c}: ''

CjkCompatibilityIdeographDash2f88d

\u{2f88d}: ''

CjkCompatibilityIdeographDash2f88e

\u{2f88e}: ''

CjkCompatibilityIdeographDash2f88f

\u{2f88f}: ''

CjkCompatibilityIdeographDash2f890

\u{2f890}: ''

CjkCompatibilityIdeographDash2f891

\u{2f891}: ''

CjkCompatibilityIdeographDash2f892

\u{2f892}: ''

CjkCompatibilityIdeographDash2f893

\u{2f893}: ''

CjkCompatibilityIdeographDash2f894

\u{2f894}: ''

CjkCompatibilityIdeographDash2f895

\u{2f895}: ''

CjkCompatibilityIdeographDash2f896

\u{2f896}: ''

CjkCompatibilityIdeographDash2f897

\u{2f897}: ''

CjkCompatibilityIdeographDash2f898

\u{2f898}: ''

CjkCompatibilityIdeographDash2f899

\u{2f899}: ''

CjkCompatibilityIdeographDash2f89a

\u{2f89a}: ''

CjkCompatibilityIdeographDash2f89b

\u{2f89b}: ''

CjkCompatibilityIdeographDash2f89c

\u{2f89c}: ''

CjkCompatibilityIdeographDash2f89d

\u{2f89d}: ''

CjkCompatibilityIdeographDash2f89e

\u{2f89e}: ''

CjkCompatibilityIdeographDash2f89f

\u{2f89f}: ''

CjkCompatibilityIdeographDash2f8a0

\u{2f8a0}: ''

CjkCompatibilityIdeographDash2f8a1

\u{2f8a1}: ''

CjkCompatibilityIdeographDash2f8a2

\u{2f8a2}: ''

CjkCompatibilityIdeographDash2f8a3

\u{2f8a3}: ''

CjkCompatibilityIdeographDash2f8a4

\u{2f8a4}: ''

CjkCompatibilityIdeographDash2f8a5

\u{2f8a5}: ''

CjkCompatibilityIdeographDash2f8a6

\u{2f8a6}: ''

CjkCompatibilityIdeographDash2f8a7

\u{2f8a7}: ''

CjkCompatibilityIdeographDash2f8a8

\u{2f8a8}: ''

CjkCompatibilityIdeographDash2f8a9

\u{2f8a9}: ''

CjkCompatibilityIdeographDash2f8aa

\u{2f8aa}: ''

CjkCompatibilityIdeographDash2f8ab

\u{2f8ab}: ''

CjkCompatibilityIdeographDash2f8ac

\u{2f8ac}: ''

CjkCompatibilityIdeographDash2f8ad

\u{2f8ad}: ''

CjkCompatibilityIdeographDash2f8ae

\u{2f8ae}: ''

CjkCompatibilityIdeographDash2f8af

\u{2f8af}: ''

CjkCompatibilityIdeographDash2f8b0

\u{2f8b0}: ''

CjkCompatibilityIdeographDash2f8b1

\u{2f8b1}: ''

CjkCompatibilityIdeographDash2f8b2

\u{2f8b2}: ''

CjkCompatibilityIdeographDash2f8b3

\u{2f8b3}: ''

CjkCompatibilityIdeographDash2f8b4

\u{2f8b4}: ''

CjkCompatibilityIdeographDash2f8b5

\u{2f8b5}: ''

CjkCompatibilityIdeographDash2f8b6

\u{2f8b6}: ''

CjkCompatibilityIdeographDash2f8b7

\u{2f8b7}: ''

CjkCompatibilityIdeographDash2f8b8

\u{2f8b8}: ''

CjkCompatibilityIdeographDash2f8b9

\u{2f8b9}: ''

CjkCompatibilityIdeographDash2f8ba

\u{2f8ba}: ''

CjkCompatibilityIdeographDash2f8bb

\u{2f8bb}: ''

CjkCompatibilityIdeographDash2f8bc

\u{2f8bc}: ''

CjkCompatibilityIdeographDash2f8bd

\u{2f8bd}: ''

CjkCompatibilityIdeographDash2f8be

\u{2f8be}: ''

CjkCompatibilityIdeographDash2f8bf

\u{2f8bf}: ''

CjkCompatibilityIdeographDash2f8c0

\u{2f8c0}: ''

CjkCompatibilityIdeographDash2f8c1

\u{2f8c1}: ''

CjkCompatibilityIdeographDash2f8c2

\u{2f8c2}: ''

CjkCompatibilityIdeographDash2f8c3

\u{2f8c3}: ''

CjkCompatibilityIdeographDash2f8c4

\u{2f8c4}: ''

CjkCompatibilityIdeographDash2f8c5

\u{2f8c5}: ''

CjkCompatibilityIdeographDash2f8c6

\u{2f8c6}: ''

CjkCompatibilityIdeographDash2f8c7

\u{2f8c7}: ''

CjkCompatibilityIdeographDash2f8c8

\u{2f8c8}: ''

CjkCompatibilityIdeographDash2f8c9

\u{2f8c9}: ''

CjkCompatibilityIdeographDash2f8ca

\u{2f8ca}: ''

CjkCompatibilityIdeographDash2f8cb

\u{2f8cb}: ''

CjkCompatibilityIdeographDash2f8cc

\u{2f8cc}: ''

CjkCompatibilityIdeographDash2f8cd

\u{2f8cd}: ''

CjkCompatibilityIdeographDash2f8ce

\u{2f8ce}: ''

CjkCompatibilityIdeographDash2f8cf

\u{2f8cf}: ''

CjkCompatibilityIdeographDash2f8d0

\u{2f8d0}: ''

CjkCompatibilityIdeographDash2f8d1

\u{2f8d1}: ''

CjkCompatibilityIdeographDash2f8d2

\u{2f8d2}: ''

CjkCompatibilityIdeographDash2f8d3

\u{2f8d3}: ''

CjkCompatibilityIdeographDash2f8d4

\u{2f8d4}: ''

CjkCompatibilityIdeographDash2f8d5

\u{2f8d5}: ''

CjkCompatibilityIdeographDash2f8d6

\u{2f8d6}: ''

CjkCompatibilityIdeographDash2f8d7

\u{2f8d7}: ''

CjkCompatibilityIdeographDash2f8d8

\u{2f8d8}: ''

CjkCompatibilityIdeographDash2f8d9

\u{2f8d9}: ''

CjkCompatibilityIdeographDash2f8da

\u{2f8da}: ''

CjkCompatibilityIdeographDash2f8db

\u{2f8db}: ''

CjkCompatibilityIdeographDash2f8dc

\u{2f8dc}: ''

CjkCompatibilityIdeographDash2f8dd

\u{2f8dd}: ''

CjkCompatibilityIdeographDash2f8de

\u{2f8de}: ''

CjkCompatibilityIdeographDash2f8df

\u{2f8df}: ''

CjkCompatibilityIdeographDash2f8e0

\u{2f8e0}: ''

CjkCompatibilityIdeographDash2f8e1

\u{2f8e1}: ''

CjkCompatibilityIdeographDash2f8e2

\u{2f8e2}: ''

CjkCompatibilityIdeographDash2f8e3

\u{2f8e3}: ''

CjkCompatibilityIdeographDash2f8e4

\u{2f8e4}: ''

CjkCompatibilityIdeographDash2f8e5

\u{2f8e5}: ''

CjkCompatibilityIdeographDash2f8e6

\u{2f8e6}: ''

CjkCompatibilityIdeographDash2f8e7

\u{2f8e7}: ''

CjkCompatibilityIdeographDash2f8e8

\u{2f8e8}: ''

CjkCompatibilityIdeographDash2f8e9

\u{2f8e9}: ''

CjkCompatibilityIdeographDash2f8ea

\u{2f8ea}: ''

CjkCompatibilityIdeographDash2f8eb

\u{2f8eb}: ''

CjkCompatibilityIdeographDash2f8ec

\u{2f8ec}: ''

CjkCompatibilityIdeographDash2f8ed

\u{2f8ed}: ''

CjkCompatibilityIdeographDash2f8ee

\u{2f8ee}: ''

CjkCompatibilityIdeographDash2f8ef

\u{2f8ef}: ''

CjkCompatibilityIdeographDash2f8f0

\u{2f8f0}: ''

CjkCompatibilityIdeographDash2f8f1

\u{2f8f1}: ''

CjkCompatibilityIdeographDash2f8f2

\u{2f8f2}: ''

CjkCompatibilityIdeographDash2f8f3

\u{2f8f3}: ''

CjkCompatibilityIdeographDash2f8f4

\u{2f8f4}: ''

CjkCompatibilityIdeographDash2f8f5

\u{2f8f5}: ''

CjkCompatibilityIdeographDash2f8f6

\u{2f8f6}: ''

CjkCompatibilityIdeographDash2f8f7

\u{2f8f7}: ''

CjkCompatibilityIdeographDash2f8f8

\u{2f8f8}: ''

CjkCompatibilityIdeographDash2f8f9

\u{2f8f9}: ''

CjkCompatibilityIdeographDash2f8fa

\u{2f8fa}: ''

CjkCompatibilityIdeographDash2f8fb

\u{2f8fb}: ''

CjkCompatibilityIdeographDash2f8fc

\u{2f8fc}: ''

CjkCompatibilityIdeographDash2f8fd

\u{2f8fd}: ''

CjkCompatibilityIdeographDash2f8fe

\u{2f8fe}: ''

CjkCompatibilityIdeographDash2f8ff

\u{2f8ff}: ''

CjkCompatibilityIdeographDash2f900

\u{2f900}: ''

CjkCompatibilityIdeographDash2f901

\u{2f901}: ''

CjkCompatibilityIdeographDash2f902

\u{2f902}: ''

CjkCompatibilityIdeographDash2f903

\u{2f903}: ''

CjkCompatibilityIdeographDash2f904

\u{2f904}: ''

CjkCompatibilityIdeographDash2f905

\u{2f905}: ''

CjkCompatibilityIdeographDash2f906

\u{2f906}: ''

CjkCompatibilityIdeographDash2f907

\u{2f907}: ''

CjkCompatibilityIdeographDash2f908

\u{2f908}: ''

CjkCompatibilityIdeographDash2f909

\u{2f909}: ''

CjkCompatibilityIdeographDash2f90a

\u{2f90a}: ''

CjkCompatibilityIdeographDash2f90b

\u{2f90b}: ''

CjkCompatibilityIdeographDash2f90c

\u{2f90c}: ''

CjkCompatibilityIdeographDash2f90d

\u{2f90d}: ''

CjkCompatibilityIdeographDash2f90e

\u{2f90e}: ''

CjkCompatibilityIdeographDash2f90f

\u{2f90f}: ''

CjkCompatibilityIdeographDash2f910

\u{2f910}: ''

CjkCompatibilityIdeographDash2f911

\u{2f911}: ''

CjkCompatibilityIdeographDash2f912

\u{2f912}: ''

CjkCompatibilityIdeographDash2f913

\u{2f913}: ''

CjkCompatibilityIdeographDash2f914

\u{2f914}: ''

CjkCompatibilityIdeographDash2f915

\u{2f915}: ''

CjkCompatibilityIdeographDash2f916

\u{2f916}: ''

CjkCompatibilityIdeographDash2f917

\u{2f917}: ''

CjkCompatibilityIdeographDash2f918

\u{2f918}: ''

CjkCompatibilityIdeographDash2f919

\u{2f919}: ''

CjkCompatibilityIdeographDash2f91a

\u{2f91a}: ''

CjkCompatibilityIdeographDash2f91b

\u{2f91b}: ''

CjkCompatibilityIdeographDash2f91c

\u{2f91c}: ''

CjkCompatibilityIdeographDash2f91d

\u{2f91d}: ''

CjkCompatibilityIdeographDash2f91e

\u{2f91e}: ''

CjkCompatibilityIdeographDash2f91f

\u{2f91f}: ''

CjkCompatibilityIdeographDash2f920

\u{2f920}: ''

CjkCompatibilityIdeographDash2f921

\u{2f921}: ''

CjkCompatibilityIdeographDash2f922

\u{2f922}: ''

CjkCompatibilityIdeographDash2f923

\u{2f923}: ''

CjkCompatibilityIdeographDash2f924

\u{2f924}: ''

CjkCompatibilityIdeographDash2f925

\u{2f925}: ''

CjkCompatibilityIdeographDash2f926

\u{2f926}: ''

CjkCompatibilityIdeographDash2f927

\u{2f927}: ''

CjkCompatibilityIdeographDash2f928

\u{2f928}: ''

CjkCompatibilityIdeographDash2f929

\u{2f929}: ''

CjkCompatibilityIdeographDash2f92a

\u{2f92a}: ''

CjkCompatibilityIdeographDash2f92b

\u{2f92b}: ''

CjkCompatibilityIdeographDash2f92c

\u{2f92c}: ''

CjkCompatibilityIdeographDash2f92d

\u{2f92d}: ''

CjkCompatibilityIdeographDash2f92e

\u{2f92e}: ''

CjkCompatibilityIdeographDash2f92f

\u{2f92f}: ''

CjkCompatibilityIdeographDash2f930

\u{2f930}: ''

CjkCompatibilityIdeographDash2f931

\u{2f931}: ''

CjkCompatibilityIdeographDash2f932

\u{2f932}: ''

CjkCompatibilityIdeographDash2f933

\u{2f933}: ''

CjkCompatibilityIdeographDash2f934

\u{2f934}: ''

CjkCompatibilityIdeographDash2f935

\u{2f935}: ''

CjkCompatibilityIdeographDash2f936

\u{2f936}: ''

CjkCompatibilityIdeographDash2f937

\u{2f937}: ''

CjkCompatibilityIdeographDash2f938

\u{2f938}: ''

CjkCompatibilityIdeographDash2f939

\u{2f939}: ''

CjkCompatibilityIdeographDash2f93a

\u{2f93a}: ''

CjkCompatibilityIdeographDash2f93b

\u{2f93b}: ''

CjkCompatibilityIdeographDash2f93c

\u{2f93c}: ''

CjkCompatibilityIdeographDash2f93d

\u{2f93d}: ''

CjkCompatibilityIdeographDash2f93e

\u{2f93e}: ''

CjkCompatibilityIdeographDash2f93f

\u{2f93f}: ''

CjkCompatibilityIdeographDash2f940

\u{2f940}: ''

CjkCompatibilityIdeographDash2f941

\u{2f941}: ''

CjkCompatibilityIdeographDash2f942

\u{2f942}: ''

CjkCompatibilityIdeographDash2f943

\u{2f943}: ''

CjkCompatibilityIdeographDash2f944

\u{2f944}: ''

CjkCompatibilityIdeographDash2f945

\u{2f945}: ''

CjkCompatibilityIdeographDash2f946

\u{2f946}: ''

CjkCompatibilityIdeographDash2f947

\u{2f947}: ''

CjkCompatibilityIdeographDash2f948

\u{2f948}: ''

CjkCompatibilityIdeographDash2f949

\u{2f949}: ''

CjkCompatibilityIdeographDash2f94a

\u{2f94a}: ''

CjkCompatibilityIdeographDash2f94b

\u{2f94b}: ''

CjkCompatibilityIdeographDash2f94c

\u{2f94c}: ''

CjkCompatibilityIdeographDash2f94d

\u{2f94d}: ''

CjkCompatibilityIdeographDash2f94e

\u{2f94e}: ''

CjkCompatibilityIdeographDash2f94f

\u{2f94f}: ''

CjkCompatibilityIdeographDash2f950

\u{2f950}: ''

CjkCompatibilityIdeographDash2f951

\u{2f951}: ''

CjkCompatibilityIdeographDash2f952

\u{2f952}: ''

CjkCompatibilityIdeographDash2f953

\u{2f953}: ''

CjkCompatibilityIdeographDash2f954

\u{2f954}: ''

CjkCompatibilityIdeographDash2f955

\u{2f955}: ''

CjkCompatibilityIdeographDash2f956

\u{2f956}: ''

CjkCompatibilityIdeographDash2f957

\u{2f957}: ''

CjkCompatibilityIdeographDash2f958

\u{2f958}: ''

CjkCompatibilityIdeographDash2f959

\u{2f959}: ''

CjkCompatibilityIdeographDash2f95a

\u{2f95a}: ''

CjkCompatibilityIdeographDash2f95b

\u{2f95b}: ''

CjkCompatibilityIdeographDash2f95c

\u{2f95c}: ''

CjkCompatibilityIdeographDash2f95d

\u{2f95d}: ''

CjkCompatibilityIdeographDash2f95e

\u{2f95e}: ''

CjkCompatibilityIdeographDash2f95f

\u{2f95f}: ''

CjkCompatibilityIdeographDash2f960

\u{2f960}: ''

CjkCompatibilityIdeographDash2f961

\u{2f961}: ''

CjkCompatibilityIdeographDash2f962

\u{2f962}: ''

CjkCompatibilityIdeographDash2f963

\u{2f963}: ''

CjkCompatibilityIdeographDash2f964

\u{2f964}: ''

CjkCompatibilityIdeographDash2f965

\u{2f965}: ''

CjkCompatibilityIdeographDash2f966

\u{2f966}: ''

CjkCompatibilityIdeographDash2f967

\u{2f967}: ''

CjkCompatibilityIdeographDash2f968

\u{2f968}: ''

CjkCompatibilityIdeographDash2f969

\u{2f969}: ''

CjkCompatibilityIdeographDash2f96a

\u{2f96a}: ''

CjkCompatibilityIdeographDash2f96b

\u{2f96b}: ''

CjkCompatibilityIdeographDash2f96c

\u{2f96c}: ''

CjkCompatibilityIdeographDash2f96d

\u{2f96d}: ''

CjkCompatibilityIdeographDash2f96e

\u{2f96e}: ''

CjkCompatibilityIdeographDash2f96f

\u{2f96f}: ''

CjkCompatibilityIdeographDash2f970

\u{2f970}: ''

CjkCompatibilityIdeographDash2f971

\u{2f971}: ''

CjkCompatibilityIdeographDash2f972

\u{2f972}: ''

CjkCompatibilityIdeographDash2f973

\u{2f973}: ''

CjkCompatibilityIdeographDash2f974

\u{2f974}: ''

CjkCompatibilityIdeographDash2f975

\u{2f975}: ''

CjkCompatibilityIdeographDash2f976

\u{2f976}: ''

CjkCompatibilityIdeographDash2f977

\u{2f977}: ''

CjkCompatibilityIdeographDash2f978

\u{2f978}: ''

CjkCompatibilityIdeographDash2f979

\u{2f979}: ''

CjkCompatibilityIdeographDash2f97a

\u{2f97a}: ''

CjkCompatibilityIdeographDash2f97b

\u{2f97b}: ''

CjkCompatibilityIdeographDash2f97c

\u{2f97c}: ''

CjkCompatibilityIdeographDash2f97d

\u{2f97d}: ''

CjkCompatibilityIdeographDash2f97e

\u{2f97e}: ''

CjkCompatibilityIdeographDash2f97f

\u{2f97f}: ''

CjkCompatibilityIdeographDash2f980

\u{2f980}: ''

CjkCompatibilityIdeographDash2f981

\u{2f981}: ''

CjkCompatibilityIdeographDash2f982

\u{2f982}: ''

CjkCompatibilityIdeographDash2f983

\u{2f983}: ''

CjkCompatibilityIdeographDash2f984

\u{2f984}: ''

CjkCompatibilityIdeographDash2f985

\u{2f985}: ''

CjkCompatibilityIdeographDash2f986

\u{2f986}: ''

CjkCompatibilityIdeographDash2f987

\u{2f987}: ''

CjkCompatibilityIdeographDash2f988

\u{2f988}: ''

CjkCompatibilityIdeographDash2f989

\u{2f989}: ''

CjkCompatibilityIdeographDash2f98a

\u{2f98a}: ''

CjkCompatibilityIdeographDash2f98b

\u{2f98b}: ''

CjkCompatibilityIdeographDash2f98c

\u{2f98c}: ''

CjkCompatibilityIdeographDash2f98d

\u{2f98d}: ''

CjkCompatibilityIdeographDash2f98e

\u{2f98e}: ''

CjkCompatibilityIdeographDash2f98f

\u{2f98f}: ''

CjkCompatibilityIdeographDash2f990

\u{2f990}: ''

CjkCompatibilityIdeographDash2f991

\u{2f991}: ''

CjkCompatibilityIdeographDash2f992

\u{2f992}: ''

CjkCompatibilityIdeographDash2f993

\u{2f993}: ''

CjkCompatibilityIdeographDash2f994

\u{2f994}: ''

CjkCompatibilityIdeographDash2f995

\u{2f995}: ''

CjkCompatibilityIdeographDash2f996

\u{2f996}: ''

CjkCompatibilityIdeographDash2f997

\u{2f997}: ''

CjkCompatibilityIdeographDash2f998

\u{2f998}: ''

CjkCompatibilityIdeographDash2f999

\u{2f999}: ''

CjkCompatibilityIdeographDash2f99a

\u{2f99a}: ''

CjkCompatibilityIdeographDash2f99b

\u{2f99b}: ''

CjkCompatibilityIdeographDash2f99c

\u{2f99c}: ''

CjkCompatibilityIdeographDash2f99d

\u{2f99d}: ''

CjkCompatibilityIdeographDash2f99e

\u{2f99e}: ''

CjkCompatibilityIdeographDash2f99f

\u{2f99f}: ''

CjkCompatibilityIdeographDash2f9a0

\u{2f9a0}: ''

CjkCompatibilityIdeographDash2f9a1

\u{2f9a1}: ''

CjkCompatibilityIdeographDash2f9a2

\u{2f9a2}: ''

CjkCompatibilityIdeographDash2f9a3

\u{2f9a3}: ''

CjkCompatibilityIdeographDash2f9a4

\u{2f9a4}: ''

CjkCompatibilityIdeographDash2f9a5

\u{2f9a5}: ''

CjkCompatibilityIdeographDash2f9a6

\u{2f9a6}: ''

CjkCompatibilityIdeographDash2f9a7

\u{2f9a7}: ''

CjkCompatibilityIdeographDash2f9a8

\u{2f9a8}: ''

CjkCompatibilityIdeographDash2f9a9

\u{2f9a9}: ''

CjkCompatibilityIdeographDash2f9aa

\u{2f9aa}: ''

CjkCompatibilityIdeographDash2f9ab

\u{2f9ab}: ''

CjkCompatibilityIdeographDash2f9ac

\u{2f9ac}: ''

CjkCompatibilityIdeographDash2f9ad

\u{2f9ad}: ''

CjkCompatibilityIdeographDash2f9ae

\u{2f9ae}: ''

CjkCompatibilityIdeographDash2f9af

\u{2f9af}: ''

CjkCompatibilityIdeographDash2f9b0

\u{2f9b0}: ''

CjkCompatibilityIdeographDash2f9b1

\u{2f9b1}: ''

CjkCompatibilityIdeographDash2f9b2

\u{2f9b2}: ''

CjkCompatibilityIdeographDash2f9b3

\u{2f9b3}: ''

CjkCompatibilityIdeographDash2f9b4

\u{2f9b4}: ''

CjkCompatibilityIdeographDash2f9b5

\u{2f9b5}: ''

CjkCompatibilityIdeographDash2f9b6

\u{2f9b6}: ''

CjkCompatibilityIdeographDash2f9b7

\u{2f9b7}: ''

CjkCompatibilityIdeographDash2f9b8

\u{2f9b8}: ''

CjkCompatibilityIdeographDash2f9b9

\u{2f9b9}: ''

CjkCompatibilityIdeographDash2f9ba

\u{2f9ba}: ''

CjkCompatibilityIdeographDash2f9bb

\u{2f9bb}: ''

CjkCompatibilityIdeographDash2f9bc

\u{2f9bc}: ''

CjkCompatibilityIdeographDash2f9bd

\u{2f9bd}: ''

CjkCompatibilityIdeographDash2f9be

\u{2f9be}: ''

CjkCompatibilityIdeographDash2f9bf

\u{2f9bf}: ''

CjkCompatibilityIdeographDash2f9c0

\u{2f9c0}: ''

CjkCompatibilityIdeographDash2f9c1

\u{2f9c1}: ''

CjkCompatibilityIdeographDash2f9c2

\u{2f9c2}: ''

CjkCompatibilityIdeographDash2f9c3

\u{2f9c3}: ''

CjkCompatibilityIdeographDash2f9c4

\u{2f9c4}: ''

CjkCompatibilityIdeographDash2f9c5

\u{2f9c5}: ''

CjkCompatibilityIdeographDash2f9c6

\u{2f9c6}: ''

CjkCompatibilityIdeographDash2f9c7

\u{2f9c7}: ''

CjkCompatibilityIdeographDash2f9c8

\u{2f9c8}: ''

CjkCompatibilityIdeographDash2f9c9

\u{2f9c9}: ''

CjkCompatibilityIdeographDash2f9ca

\u{2f9ca}: ''

CjkCompatibilityIdeographDash2f9cb

\u{2f9cb}: ''

CjkCompatibilityIdeographDash2f9cc

\u{2f9cc}: ''

CjkCompatibilityIdeographDash2f9cd

\u{2f9cd}: ''

CjkCompatibilityIdeographDash2f9ce

\u{2f9ce}: ''

CjkCompatibilityIdeographDash2f9cf

\u{2f9cf}: ''

CjkCompatibilityIdeographDash2f9d0

\u{2f9d0}: ''

CjkCompatibilityIdeographDash2f9d1

\u{2f9d1}: ''

CjkCompatibilityIdeographDash2f9d2

\u{2f9d2}: ''

CjkCompatibilityIdeographDash2f9d3

\u{2f9d3}: ''

CjkCompatibilityIdeographDash2f9d4

\u{2f9d4}: ''

CjkCompatibilityIdeographDash2f9d5

\u{2f9d5}: ''

CjkCompatibilityIdeographDash2f9d6

\u{2f9d6}: ''

CjkCompatibilityIdeographDash2f9d7

\u{2f9d7}: ''

CjkCompatibilityIdeographDash2f9d8

\u{2f9d8}: ''

CjkCompatibilityIdeographDash2f9d9

\u{2f9d9}: ''

CjkCompatibilityIdeographDash2f9da

\u{2f9da}: ''

CjkCompatibilityIdeographDash2f9db

\u{2f9db}: ''

CjkCompatibilityIdeographDash2f9dc

\u{2f9dc}: ''

CjkCompatibilityIdeographDash2f9dd

\u{2f9dd}: ''

CjkCompatibilityIdeographDash2f9de

\u{2f9de}: ''

CjkCompatibilityIdeographDash2f9df

\u{2f9df}: ''

CjkCompatibilityIdeographDash2f9e0

\u{2f9e0}: ''

CjkCompatibilityIdeographDash2f9e1

\u{2f9e1}: ''

CjkCompatibilityIdeographDash2f9e2

\u{2f9e2}: ''

CjkCompatibilityIdeographDash2f9e3

\u{2f9e3}: ''

CjkCompatibilityIdeographDash2f9e4

\u{2f9e4}: ''

CjkCompatibilityIdeographDash2f9e5

\u{2f9e5}: ''

CjkCompatibilityIdeographDash2f9e6

\u{2f9e6}: ''

CjkCompatibilityIdeographDash2f9e7

\u{2f9e7}: ''

CjkCompatibilityIdeographDash2f9e8

\u{2f9e8}: ''

CjkCompatibilityIdeographDash2f9e9

\u{2f9e9}: ''

CjkCompatibilityIdeographDash2f9ea

\u{2f9ea}: ''

CjkCompatibilityIdeographDash2f9eb

\u{2f9eb}: ''

CjkCompatibilityIdeographDash2f9ec

\u{2f9ec}: ''

CjkCompatibilityIdeographDash2f9ed

\u{2f9ed}: ''

CjkCompatibilityIdeographDash2f9ee

\u{2f9ee}: ''

CjkCompatibilityIdeographDash2f9ef

\u{2f9ef}: ''

CjkCompatibilityIdeographDash2f9f0

\u{2f9f0}: ''

CjkCompatibilityIdeographDash2f9f1

\u{2f9f1}: ''

CjkCompatibilityIdeographDash2f9f2

\u{2f9f2}: ''

CjkCompatibilityIdeographDash2f9f3

\u{2f9f3}: ''

CjkCompatibilityIdeographDash2f9f4

\u{2f9f4}: ''

CjkCompatibilityIdeographDash2f9f5

\u{2f9f5}: ''

CjkCompatibilityIdeographDash2f9f6

\u{2f9f6}: ''

CjkCompatibilityIdeographDash2f9f7

\u{2f9f7}: ''

CjkCompatibilityIdeographDash2f9f8

\u{2f9f8}: ''

CjkCompatibilityIdeographDash2f9f9

\u{2f9f9}: ''

CjkCompatibilityIdeographDash2f9fa

\u{2f9fa}: ''

CjkCompatibilityIdeographDash2f9fb

\u{2f9fb}: ''

CjkCompatibilityIdeographDash2f9fc

\u{2f9fc}: ''

CjkCompatibilityIdeographDash2f9fd

\u{2f9fd}: ''

CjkCompatibilityIdeographDash2f9fe

\u{2f9fe}: ''

CjkCompatibilityIdeographDash2f9ff

\u{2f9ff}: ''

CjkCompatibilityIdeographDash2fa00

\u{2fa00}: ''

CjkCompatibilityIdeographDash2fa01

\u{2fa01}: ''

CjkCompatibilityIdeographDash2fa02

\u{2fa02}: ''

CjkCompatibilityIdeographDash2fa03

\u{2fa03}: ''

CjkCompatibilityIdeographDash2fa04

\u{2fa04}: ''

CjkCompatibilityIdeographDash2fa05

\u{2fa05}: ''

CjkCompatibilityIdeographDash2fa06

\u{2fa06}: ''

CjkCompatibilityIdeographDash2fa07

\u{2fa07}: ''

CjkCompatibilityIdeographDash2fa08

\u{2fa08}: ''

CjkCompatibilityIdeographDash2fa09

\u{2fa09}: ''

CjkCompatibilityIdeographDash2fa0a

\u{2fa0a}: ''

CjkCompatibilityIdeographDash2fa0b

\u{2fa0b}: ''

CjkCompatibilityIdeographDash2fa0c

\u{2fa0c}: ''

CjkCompatibilityIdeographDash2fa0d

\u{2fa0d}: ''

CjkCompatibilityIdeographDash2fa0e

\u{2fa0e}: ''

CjkCompatibilityIdeographDash2fa0f

\u{2fa0f}: ''

CjkCompatibilityIdeographDash2fa10

\u{2fa10}: ''

CjkCompatibilityIdeographDash2fa11

\u{2fa11}: ''

CjkCompatibilityIdeographDash2fa12

\u{2fa12}: ''

CjkCompatibilityIdeographDash2fa13

\u{2fa13}: ''

CjkCompatibilityIdeographDash2fa14

\u{2fa14}: ''

CjkCompatibilityIdeographDash2fa15

\u{2fa15}: ''

CjkCompatibilityIdeographDash2fa16

\u{2fa16}: ''

CjkCompatibilityIdeographDash2fa17

\u{2fa17}: ''

CjkCompatibilityIdeographDash2fa18

\u{2fa18}: ''

CjkCompatibilityIdeographDash2fa19

\u{2fa19}: ''

CjkCompatibilityIdeographDash2fa1a

\u{2fa1a}: ''

CjkCompatibilityIdeographDash2fa1b

\u{2fa1b}: ''

CjkCompatibilityIdeographDash2fa1c

\u{2fa1c}: ''

CjkCompatibilityIdeographDash2fa1d

\u{2fa1d}: ''

Methods

impl CJKCompatibilityIdeographsSupplement

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case