Enum unicode_types::cjk_radicals_supplement::CJKRadicalsSupplement

pub enum CJKRadicalsSupplement {
    CjkRadicalRepeat,
    CjkRadicalCliff,
    CjkRadicalSecondOne,
    CjkRadicalSecondTwo,
    CjkRadicalSecondThree,
    CjkRadicalPerson,
    CjkRadicalBox,
    CjkRadicalTable,
    CjkRadicalKnifeOne,
    CjkRadicalKnifeTwo,
    CjkRadicalDivination,
    CjkRadicalSeal,
    CjkRadicalSmallOne,
    CjkRadicalSmallTwo,
    CjkRadicalLameOne,
    CjkRadicalLameTwo,
    CjkRadicalLameThree,
    CjkRadicalLameFour,
    CjkRadicalSnake,
    CjkRadicalThread,
    CjkRadicalSnoutOne,
    CjkRadicalSnoutTwo,
    CjkRadicalHeartOne,
    CjkRadicalHeartTwo,
    CjkRadicalHand,
    CjkRadicalRap,
    CjkRadicalChoke,
    CjkRadicalSun,
    CjkRadicalMoon,
    CjkRadicalDeath,
    CjkRadicalMother,
    CjkRadicalCivilian,
    CjkRadicalWaterOne,
    CjkRadicalWaterTwo,
    CjkRadicalFire,
    CjkRadicalPawOne,
    CjkRadicalPawTwo,
    CjkRadicalSimplifiedHalfTreeTrunk,
    CjkRadicalCow,
    CjkRadicalDog,
    CjkRadicalJade,
    CjkRadicalBoltOfCloth,
    CjkRadicalEye,
    CjkRadicalSpiritOne,
    CjkRadicalSpiritTwo,
    CjkRadicalBamboo,
    CjkRadicalSilk,
    CjkRadicalCDashSimplifiedSilk,
    CjkRadicalNetOne,
    CjkRadicalNetTwo,
    CjkRadicalNetThree,
    CjkRadicalNetFour,
    CjkRadicalMesh,
    CjkRadicalSheep,
    CjkRadicalRam,
    CjkRadicalEwe,
    CjkRadicalOld,
    CjkRadicalBrushOne,
    CjkRadicalBrushTwo,
    CjkRadicalMeat,
    CjkRadicalMortar,
    CjkRadicalGrassOne,
    CjkRadicalGrassTwo,
    CjkRadicalGrassThree,
    CjkRadicalTiger,
    CjkRadicalClothes,
    CjkRadicalWestOne,
    CjkRadicalWestTwo,
    CjkRadicalCDashSimplifiedSee,
    CjkRadicalSimplifiedHorn,
    CjkRadicalHorn,
    CjkRadicalCDashSimplifiedSpeech,
    CjkRadicalCDashSimplifiedShell,
    CjkRadicalFoot,
    CjkRadicalCDashSimplifiedCart,
    CjkRadicalSimplifiedWalk,
    CjkRadicalWalkOne,
    CjkRadicalWalkTwo,
    CjkRadicalCity,
    CjkRadicalCDashSimplifiedGold,
    CjkRadicalLongOne,
    CjkRadicalLongTwo,
    CjkRadicalCDashSimplifiedLong,
    CjkRadicalCDashSimplifiedGate,
    CjkRadicalMoundOne,
    CjkRadicalMoundTwo,
    CjkRadicalRain,
    CjkRadicalBlue,
    CjkRadicalCDashSimplifiedTannedLeather,
    CjkRadicalCDashSimplifiedLeaf,
    CjkRadicalCDashSimplifiedWind,
    CjkRadicalCDashSimplifiedFly,
    CjkRadicalEatOne,
    CjkRadicalEatTwo,
    CjkRadicalEatThree,
    CjkRadicalCDashSimplifiedEat,
    CjkRadicalHead,
    CjkRadicalCDashSimplifiedHorse,
    CjkRadicalBone,
    CjkRadicalGhost,
    CjkRadicalCDashSimplifiedFish,
    CjkRadicalCDashSimplifiedBird,
    CjkRadicalCDashSimplifiedSalt,
    CjkRadicalSimplifiedWheat,
    CjkRadicalSimplifiedYellow,
    CjkRadicalCDashSimplifiedFrog,
    CjkRadicalJDashSimplifiedEven,
    CjkRadicalCDashSimplifiedEven,
    CjkRadicalJDashSimplifiedTooth,
    CjkRadicalCDashSimplifiedTooth,
    CjkRadicalJDashSimplifiedDragon,
    CjkRadicalCDashSimplifiedDragon,
    CjkRadicalTurtle,
    CjkRadicalJDashSimplifiedTurtle,
    CjkRadicalCDashSimplifiedTurtle,
}

An enum to represent all characters in the CJKRadicalsSupplement block.

Variants
CjkRadicalRepeat

\u{2e80}: '⺀'

CjkRadicalCliff

\u{2e81}: '⺁'

CjkRadicalSecondOne

\u{2e82}: '⺂'

CjkRadicalSecondTwo

\u{2e83}: '⺃'

CjkRadicalSecondThree

\u{2e84}: '⺄'

CjkRadicalPerson

\u{2e85}: '⺅'

CjkRadicalBox

\u{2e86}: '⺆'

CjkRadicalTable

\u{2e87}: '⺇'

CjkRadicalKnifeOne

\u{2e88}: '⺈'

CjkRadicalKnifeTwo

\u{2e89}: '⺉'

CjkRadicalDivination

\u{2e8a}: '⺊'

CjkRadicalSeal

\u{2e8b}: '⺋'

CjkRadicalSmallOne

\u{2e8c}: '⺌'

CjkRadicalSmallTwo

\u{2e8d}: '⺍'

CjkRadicalLameOne

\u{2e8e}: '⺎'

CjkRadicalLameTwo

\u{2e8f}: '⺏'

CjkRadicalLameThree

\u{2e90}: '⺐'

CjkRadicalLameFour

\u{2e91}: '⺑'

CjkRadicalSnake

\u{2e92}: '⺒'

CjkRadicalThread

\u{2e93}: '⺓'

CjkRadicalSnoutOne

\u{2e94}: '⺔'

CjkRadicalSnoutTwo

\u{2e95}: '⺕'

CjkRadicalHeartOne

\u{2e96}: '⺖'

CjkRadicalHeartTwo

\u{2e97}: '⺗'

CjkRadicalHand

\u{2e98}: '⺘'

CjkRadicalRap

\u{2e99}: '⺙'

CjkRadicalChoke

\u{2e9b}: '⺛'

CjkRadicalSun

\u{2e9c}: '⺜'

CjkRadicalMoon

\u{2e9d}: '⺝'

CjkRadicalDeath

\u{2e9e}: '⺞'

CjkRadicalMother

\u{2e9f}: '⺟'

CjkRadicalCivilian

\u{2ea0}: '⺠'

CjkRadicalWaterOne

\u{2ea1}: '⺡'

CjkRadicalWaterTwo

\u{2ea2}: '⺢'

CjkRadicalFire

\u{2ea3}: '⺣'

CjkRadicalPawOne

\u{2ea4}: '⺤'

CjkRadicalPawTwo

\u{2ea5}: '⺥'

CjkRadicalSimplifiedHalfTreeTrunk

\u{2ea6}: '⺦'

CjkRadicalCow

\u{2ea7}: '⺧'

CjkRadicalDog

\u{2ea8}: '⺨'

CjkRadicalJade

\u{2ea9}: '⺩'

CjkRadicalBoltOfCloth

\u{2eaa}: '⺪'

CjkRadicalEye

\u{2eab}: '⺫'

CjkRadicalSpiritOne

\u{2eac}: '⺬'

CjkRadicalSpiritTwo

\u{2ead}: '⺭'

CjkRadicalBamboo

\u{2eae}: '⺮'

CjkRadicalSilk

\u{2eaf}: '⺯'

CjkRadicalCDashSimplifiedSilk

\u{2eb0}: '⺰'

CjkRadicalNetOne

\u{2eb1}: '⺱'

CjkRadicalNetTwo

\u{2eb2}: '⺲'

CjkRadicalNetThree

\u{2eb3}: '⺳'

CjkRadicalNetFour

\u{2eb4}: '⺴'

CjkRadicalMesh

\u{2eb5}: '⺵'

CjkRadicalSheep

\u{2eb6}: '⺶'

CjkRadicalRam

\u{2eb7}: '⺷'

CjkRadicalEwe

\u{2eb8}: '⺸'

CjkRadicalOld

\u{2eb9}: '⺹'

CjkRadicalBrushOne

\u{2eba}: '⺺'

CjkRadicalBrushTwo

\u{2ebb}: '⺻'

CjkRadicalMeat

\u{2ebc}: '⺼'

CjkRadicalMortar

\u{2ebd}: '⺽'

CjkRadicalGrassOne

\u{2ebe}: '⺾'

CjkRadicalGrassTwo

\u{2ebf}: '⺿'

CjkRadicalGrassThree

\u{2ec0}: '⻀'

CjkRadicalTiger

\u{2ec1}: '⻁'

CjkRadicalClothes

\u{2ec2}: '⻂'

CjkRadicalWestOne

\u{2ec3}: '⻃'

CjkRadicalWestTwo

\u{2ec4}: '⻄'

CjkRadicalCDashSimplifiedSee

\u{2ec5}: '⻅'

CjkRadicalSimplifiedHorn

\u{2ec6}: '⻆'

CjkRadicalHorn

\u{2ec7}: '⻇'

CjkRadicalCDashSimplifiedSpeech

\u{2ec8}: '⻈'

CjkRadicalCDashSimplifiedShell

\u{2ec9}: '⻉'

CjkRadicalFoot

\u{2eca}: '⻊'

CjkRadicalCDashSimplifiedCart

\u{2ecb}: '⻋'

CjkRadicalSimplifiedWalk

\u{2ecc}: '⻌'

CjkRadicalWalkOne

\u{2ecd}: '⻍'

CjkRadicalWalkTwo

\u{2ece}: '⻎'

CjkRadicalCity

\u{2ecf}: '⻏'

CjkRadicalCDashSimplifiedGold

\u{2ed0}: '⻐'

CjkRadicalLongOne

\u{2ed1}: '⻑'

CjkRadicalLongTwo

\u{2ed2}: '⻒'

CjkRadicalCDashSimplifiedLong

\u{2ed3}: '⻓'

CjkRadicalCDashSimplifiedGate

\u{2ed4}: '⻔'

CjkRadicalMoundOne

\u{2ed5}: '⻕'

CjkRadicalMoundTwo

\u{2ed6}: '⻖'

CjkRadicalRain

\u{2ed7}: '⻗'

CjkRadicalBlue

\u{2ed8}: '⻘'

CjkRadicalCDashSimplifiedTannedLeather

\u{2ed9}: '⻙'

CjkRadicalCDashSimplifiedLeaf

\u{2eda}: '⻚'

CjkRadicalCDashSimplifiedWind

\u{2edb}: '⻛'

CjkRadicalCDashSimplifiedFly

\u{2edc}: '⻜'

CjkRadicalEatOne

\u{2edd}: '⻝'

CjkRadicalEatTwo

\u{2ede}: '⻞'

CjkRadicalEatThree

\u{2edf}: '⻟'

CjkRadicalCDashSimplifiedEat

\u{2ee0}: '⻠'

CjkRadicalHead

\u{2ee1}: '⻡'

CjkRadicalCDashSimplifiedHorse

\u{2ee2}: '⻢'

CjkRadicalBone

\u{2ee3}: '⻣'

CjkRadicalGhost

\u{2ee4}: '⻤'

CjkRadicalCDashSimplifiedFish

\u{2ee5}: '⻥'

CjkRadicalCDashSimplifiedBird

\u{2ee6}: '⻦'

CjkRadicalCDashSimplifiedSalt

\u{2ee7}: '⻧'

CjkRadicalSimplifiedWheat

\u{2ee8}: '⻨'

CjkRadicalSimplifiedYellow

\u{2ee9}: '⻩'

CjkRadicalCDashSimplifiedFrog

\u{2eea}: '⻪'

CjkRadicalJDashSimplifiedEven

\u{2eeb}: '⻫'

CjkRadicalCDashSimplifiedEven

\u{2eec}: '⻬'

CjkRadicalJDashSimplifiedTooth

\u{2eed}: '⻭'

CjkRadicalCDashSimplifiedTooth

\u{2eee}: '⻮'

CjkRadicalJDashSimplifiedDragon

\u{2eef}: '⻯'

CjkRadicalCDashSimplifiedDragon

\u{2ef0}: '⻰'

CjkRadicalTurtle

\u{2ef1}: '⻱'

CjkRadicalJDashSimplifiedTurtle

\u{2ef2}: '⻲'

CjkRadicalCDashSimplifiedTurtle

\u{2ef3}: '⻳'

Methods

impl CJKRadicalsSupplement

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case