Enum unicode_types::cjk_symbols_and_punctuation::CJKSymbolsandPunctuation

pub enum CJKSymbolsandPunctuation {
    IdeographicSpace,
    IdeographicComma,
    IdeographicFullStop,
    DittoMark,
    JapaneseIndustrialStandardSymbol,
    IdeographicIterationMark,
    IdeographicClosingMark,
    IdeographicNumberZero,
    LeftAngleBracket,
    RightAngleBracket,
    LeftDoubleAngleBracket,
    RightDoubleAngleBracket,
    LeftCornerBracket,
    RightCornerBracket,
    LeftWhiteCornerBracket,
    RightWhiteCornerBracket,
    LeftBlackLenticularBracket,
    RightBlackLenticularBracket,
    PostalMark,
    GetaMark,
    LeftTortoiseShellBracket,
    RightTortoiseShellBracket,
    LeftWhiteLenticularBracket,
    RightWhiteLenticularBracket,
    LeftWhiteTortoiseShellBracket,
    RightWhiteTortoiseShellBracket,
    LeftWhiteSquareBracket,
    RightWhiteSquareBracket,
    WaveDash,
    ReversedDoublePrimeQuotationMark,
    DoublePrimeQuotationMark,
    LowDoublePrimeQuotationMark,
    PostalMarkFace,
    HangzhouNumeralOne,
    HangzhouNumeralTwo,
    HangzhouNumeralThree,
    HangzhouNumeralFour,
    HangzhouNumeralFive,
    HangzhouNumeralSix,
    HangzhouNumeralSeven,
    HangzhouNumeralEight,
    HangzhouNumeralNine,
    IdeographicLevelToneMark,
    IdeographicRisingToneMark,
    IdeographicDepartingToneMark,
    IdeographicEnteringToneMark,
    HangulSingleDotToneMark,
    HangulDoubleDotToneMark,
    WavyDash,
    VerticalKanaRepeatMark,
    VerticalKanaRepeatWithVoicedSoundMark,
    VerticalKanaRepeatMarkUpperHalf,
    VerticalKanaRepeatWithVoicedSoundMarkUpperHalf,
    VerticalKanaRepeatMarkLowerHalf,
    CircledPostalMark,
    IdeographicTelegraphLineFeedSeparatorSymbol,
    HangzhouNumeralTen,
    HangzhouNumeralTwenty,
    HangzhouNumeralThirty,
    VerticalIdeographicIterationMark,
    MasuMark,
    PartAlternationMark,
    IdeographicVariationIndicator,
}

An enum to represent all characters in the CJKSymbolsandPunctuation block.

Variants
IdeographicSpace

\u{3000}: ' '

IdeographicComma

\u{3001}: '、'

IdeographicFullStop

\u{3002}: '。'

DittoMark

\u{3003}: '〃'

JapaneseIndustrialStandardSymbol

\u{3004}: '〄'

IdeographicIterationMark

\u{3005}: '々'

IdeographicClosingMark

\u{3006}: '〆'

IdeographicNumberZero

\u{3007}: '〇'

LeftAngleBracket

\u{3008}: '〈'

RightAngleBracket

\u{3009}: '〉'

LeftDoubleAngleBracket

\u{300a}: '《'

RightDoubleAngleBracket

\u{300b}: '》'

LeftCornerBracket

\u{300c}: '「'

RightCornerBracket

\u{300d}: '」'

LeftWhiteCornerBracket

\u{300e}: '『'

RightWhiteCornerBracket

\u{300f}: '』'

LeftBlackLenticularBracket

\u{3010}: '【'

RightBlackLenticularBracket

\u{3011}: '】'

PostalMark

\u{3012}: '〒'

GetaMark

\u{3013}: '〓'

LeftTortoiseShellBracket

\u{3014}: '〔'

RightTortoiseShellBracket

\u{3015}: '〕'

LeftWhiteLenticularBracket

\u{3016}: '〖'

RightWhiteLenticularBracket

\u{3017}: '〗'

LeftWhiteTortoiseShellBracket

\u{3018}: '〘'

RightWhiteTortoiseShellBracket

\u{3019}: '〙'

LeftWhiteSquareBracket

\u{301a}: '〚'

RightWhiteSquareBracket

\u{301b}: '〛'

WaveDash

\u{301c}: '〜'

ReversedDoublePrimeQuotationMark

\u{301d}: '〝'

DoublePrimeQuotationMark

\u{301e}: '〞'

LowDoublePrimeQuotationMark

\u{301f}: '〟'

PostalMarkFace

\u{3020}: '〠'

HangzhouNumeralOne

\u{3021}: '〡'

HangzhouNumeralTwo

\u{3022}: '〢'

HangzhouNumeralThree

\u{3023}: '〣'

HangzhouNumeralFour

\u{3024}: '〤'

HangzhouNumeralFive

\u{3025}: '〥'

HangzhouNumeralSix

\u{3026}: '〦'

HangzhouNumeralSeven

\u{3027}: '〧'

HangzhouNumeralEight

\u{3028}: '〨'

HangzhouNumeralNine

\u{3029}: '〩'

IdeographicLevelToneMark

\u{302a}: '〪'

IdeographicRisingToneMark

\u{302b}: '〫'

IdeographicDepartingToneMark

\u{302c}: '〬'

IdeographicEnteringToneMark

\u{302d}: '〭'

HangulSingleDotToneMark

\u{302e}: '〮'

HangulDoubleDotToneMark

\u{302f}: '〯'

WavyDash

\u{3030}: '〰'

VerticalKanaRepeatMark

\u{3031}: '〱'

VerticalKanaRepeatWithVoicedSoundMark

\u{3032}: '〲'

VerticalKanaRepeatMarkUpperHalf

\u{3033}: '〳'

VerticalKanaRepeatWithVoicedSoundMarkUpperHalf

\u{3034}: '〴'

VerticalKanaRepeatMarkLowerHalf

\u{3035}: '〵'

CircledPostalMark

\u{3036}: '〶'

IdeographicTelegraphLineFeedSeparatorSymbol

\u{3037}: '〷'

HangzhouNumeralTen

\u{3038}: '〸'

HangzhouNumeralTwenty

\u{3039}: '〹'

HangzhouNumeralThirty

\u{303a}: '〺'

VerticalIdeographicIterationMark

\u{303b}: '〻'

MasuMark

\u{303c}: '〼'

PartAlternationMark

\u{303d}: '〽'

IdeographicVariationIndicator

\u{303e}: '〾'

Methods

impl CJKSymbolsandPunctuation

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case