Enum unicode_types::coptic::Coptic

pub enum Coptic {
    CapitalLetterAlfa,
    SmallLetterAlfa,
    CapitalLetterVida,
    SmallLetterVida,
    CapitalLetterGamma,
    SmallLetterGamma,
    CapitalLetterDalda,
    SmallLetterDalda,
    CapitalLetterEie,
    SmallLetterEie,
    CapitalLetterSou,
    SmallLetterSou,
    CapitalLetterZata,
    SmallLetterZata,
    CapitalLetterHate,
    SmallLetterHate,
    CapitalLetterThethe,
    SmallLetterThethe,
    CapitalLetterIauda,
    SmallLetterIauda,
    CapitalLetterKapa,
    SmallLetterKapa,
    CapitalLetterLaula,
    SmallLetterLaula,
    CapitalLetterMi,
    SmallLetterMi,
    CapitalLetterNi,
    SmallLetterNi,
    CapitalLetterKsi,
    SmallLetterKsi,
    CapitalLetterO,
    SmallLetterO,
    CapitalLetterPi,
    SmallLetterPi,
    CapitalLetterRo,
    SmallLetterRo,
    CapitalLetterSima,
    SmallLetterSima,
    CapitalLetterTau,
    SmallLetterTau,
    CapitalLetterUa,
    SmallLetterUa,
    CapitalLetterFi,
    SmallLetterFi,
    CapitalLetterKhi,
    SmallLetterKhi,
    CapitalLetterPsi,
    SmallLetterPsi,
    CapitalLetterOou,
    SmallLetterOou,
    CapitalLetterDialectDashPAlef,
    SmallLetterDialectDashPAlef,
    CapitalLetterOldAin,
    SmallLetterOldAin,
    CapitalLetterCryptogrammicEie,
    SmallLetterCryptogrammicEie,
    CapitalLetterDialectDashPKapa,
    SmallLetterDialectDashPKapa,
    CapitalLetterDialectDashPNi,
    SmallLetterDialectDashPNi,
    CapitalLetterCryptogrammicNi,
    SmallLetterCryptogrammicNi,
    CapitalLetterOldOou,
    SmallLetterOldOou,
    CapitalLetterSampi,
    SmallLetterSampi,
    CapitalLetterCrossedShei,
    SmallLetterCrossedShei,
    CapitalLetterOldShei,
    SmallLetterOldShei,
    CapitalLetterOldEsh,
    SmallLetterOldEsh,
    CapitalLetterAkhmimicKhei,
    SmallLetterAkhmimicKhei,
    CapitalLetterDialectDashPHori,
    SmallLetterDialectDashPHori,
    CapitalLetterOldHori,
    SmallLetterOldHori,
    CapitalLetterOldHa,
    SmallLetterOldHa,
    CapitalLetterLDashShapedHa,
    SmallLetterLDashShapedHa,
    CapitalLetterOldHei,
    SmallLetterOldHei,
    CapitalLetterOldHat,
    SmallLetterOldHat,
    CapitalLetterOldGangia,
    SmallLetterOldGangia,
    CapitalLetterOldDja,
    SmallLetterOldDja,
    CapitalLetterOldShima,
    SmallLetterOldShima,
    CapitalLetterOldNubianShima,
    SmallLetterOldNubianShima,
    CapitalLetterOldNubianNgi,
    SmallLetterOldNubianNgi,
    CapitalLetterOldNubianNyi,
    SmallLetterOldNubianNyi,
    CapitalLetterOldNubianWau,
    SmallLetterOldNubianWau,
    SymbolKai,
    SymbolMiRo,
    SymbolPiRo,
    SymbolStauros,
    SymbolTauRo,
    SymbolKhiRo,
    SymbolShimaSima,
    CapitalLetterCryptogrammicShei,
    SmallLetterCryptogrammicShei,
    CapitalLetterCryptogrammicGangia,
    SmallLetterCryptogrammicGangia,
    CombiningNiAbove,
    CombiningSpiritusAsper,
    CombiningSpiritusLenis,
    CapitalLetterBohairicKhei,
    SmallLetterBohairicKhei,
    OldNubianFullStop,
    OldNubianDirectQuestionMark,
    OldNubianIndirectQuestionMark,
    OldNubianVerseDivider,
    FractionOneHalf,
    FullStop,
}

An enum to represent all characters in the Coptic block.

Variants
CapitalLetterAlfa

\u{2c80}: 'Ⲁ'

SmallLetterAlfa

\u{2c81}: 'ⲁ'

CapitalLetterVida

\u{2c82}: 'Ⲃ'

SmallLetterVida

\u{2c83}: 'ⲃ'

CapitalLetterGamma

\u{2c84}: 'Ⲅ'

SmallLetterGamma

\u{2c85}: 'ⲅ'

CapitalLetterDalda

\u{2c86}: 'Ⲇ'

SmallLetterDalda

\u{2c87}: 'ⲇ'

CapitalLetterEie

\u{2c88}: 'Ⲉ'

SmallLetterEie

\u{2c89}: 'ⲉ'

CapitalLetterSou

\u{2c8a}: 'Ⲋ'

SmallLetterSou

\u{2c8b}: 'ⲋ'

CapitalLetterZata

\u{2c8c}: 'Ⲍ'

SmallLetterZata

\u{2c8d}: 'ⲍ'

CapitalLetterHate

\u{2c8e}: 'Ⲏ'

SmallLetterHate

\u{2c8f}: 'ⲏ'

CapitalLetterThethe

\u{2c90}: 'Ⲑ'

SmallLetterThethe

\u{2c91}: 'ⲑ'

CapitalLetterIauda

\u{2c92}: 'Ⲓ'

SmallLetterIauda

\u{2c93}: 'ⲓ'

CapitalLetterKapa

\u{2c94}: 'Ⲕ'

SmallLetterKapa

\u{2c95}: 'ⲕ'

CapitalLetterLaula

\u{2c96}: 'Ⲗ'

SmallLetterLaula

\u{2c97}: 'ⲗ'

CapitalLetterMi

\u{2c98}: 'Ⲙ'

SmallLetterMi

\u{2c99}: 'ⲙ'

CapitalLetterNi

\u{2c9a}: 'Ⲛ'

SmallLetterNi

\u{2c9b}: 'ⲛ'

CapitalLetterKsi

\u{2c9c}: 'Ⲝ'

SmallLetterKsi

\u{2c9d}: 'ⲝ'

CapitalLetterO

\u{2c9e}: 'Ⲟ'

SmallLetterO

\u{2c9f}: 'ⲟ'

CapitalLetterPi

\u{2ca0}: 'Ⲡ'

SmallLetterPi

\u{2ca1}: 'ⲡ'

CapitalLetterRo

\u{2ca2}: 'Ⲣ'

SmallLetterRo

\u{2ca3}: 'ⲣ'

CapitalLetterSima

\u{2ca4}: 'Ⲥ'

SmallLetterSima

\u{2ca5}: 'ⲥ'

CapitalLetterTau

\u{2ca6}: 'Ⲧ'

SmallLetterTau

\u{2ca7}: 'ⲧ'

CapitalLetterUa

\u{2ca8}: 'Ⲩ'

SmallLetterUa

\u{2ca9}: 'ⲩ'

CapitalLetterFi

\u{2caa}: 'Ⲫ'

SmallLetterFi

\u{2cab}: 'ⲫ'

CapitalLetterKhi

\u{2cac}: 'Ⲭ'

SmallLetterKhi

\u{2cad}: 'ⲭ'

CapitalLetterPsi

\u{2cae}: 'Ⲯ'

SmallLetterPsi

\u{2caf}: 'ⲯ'

CapitalLetterOou

\u{2cb0}: 'Ⲱ'

SmallLetterOou

\u{2cb1}: 'ⲱ'

CapitalLetterDialectDashPAlef

\u{2cb2}: 'Ⲳ'

SmallLetterDialectDashPAlef

\u{2cb3}: 'ⲳ'

CapitalLetterOldAin

\u{2cb4}: 'Ⲵ'

SmallLetterOldAin

\u{2cb5}: 'ⲵ'

CapitalLetterCryptogrammicEie

\u{2cb6}: 'Ⲷ'

SmallLetterCryptogrammicEie

\u{2cb7}: 'ⲷ'

CapitalLetterDialectDashPKapa

\u{2cb8}: 'Ⲹ'

SmallLetterDialectDashPKapa

\u{2cb9}: 'ⲹ'

CapitalLetterDialectDashPNi

\u{2cba}: 'Ⲻ'

SmallLetterDialectDashPNi

\u{2cbb}: 'ⲻ'

CapitalLetterCryptogrammicNi

\u{2cbc}: 'Ⲽ'

SmallLetterCryptogrammicNi

\u{2cbd}: 'ⲽ'

CapitalLetterOldOou

\u{2cbe}: 'Ⲿ'

SmallLetterOldOou

\u{2cbf}: 'ⲿ'

CapitalLetterSampi

\u{2cc0}: 'Ⳁ'

SmallLetterSampi

\u{2cc1}: 'ⳁ'

CapitalLetterCrossedShei

\u{2cc2}: 'Ⳃ'

SmallLetterCrossedShei

\u{2cc3}: 'ⳃ'

CapitalLetterOldShei

\u{2cc4}: 'Ⳅ'

SmallLetterOldShei

\u{2cc5}: 'ⳅ'

CapitalLetterOldEsh

\u{2cc6}: 'Ⳇ'

SmallLetterOldEsh

\u{2cc7}: 'ⳇ'

CapitalLetterAkhmimicKhei

\u{2cc8}: 'Ⳉ'

SmallLetterAkhmimicKhei

\u{2cc9}: 'ⳉ'

CapitalLetterDialectDashPHori

\u{2cca}: 'Ⳋ'

SmallLetterDialectDashPHori

\u{2ccb}: 'ⳋ'

CapitalLetterOldHori

\u{2ccc}: 'Ⳍ'

SmallLetterOldHori

\u{2ccd}: 'ⳍ'

CapitalLetterOldHa

\u{2cce}: 'Ⳏ'

SmallLetterOldHa

\u{2ccf}: 'ⳏ'

CapitalLetterLDashShapedHa

\u{2cd0}: 'Ⳑ'

SmallLetterLDashShapedHa

\u{2cd1}: 'ⳑ'

CapitalLetterOldHei

\u{2cd2}: 'Ⳓ'

SmallLetterOldHei

\u{2cd3}: 'ⳓ'

CapitalLetterOldHat

\u{2cd4}: 'Ⳕ'

SmallLetterOldHat

\u{2cd5}: 'ⳕ'

CapitalLetterOldGangia

\u{2cd6}: 'Ⳗ'

SmallLetterOldGangia

\u{2cd7}: 'ⳗ'

CapitalLetterOldDja

\u{2cd8}: 'Ⳙ'

SmallLetterOldDja

\u{2cd9}: 'ⳙ'

CapitalLetterOldShima

\u{2cda}: 'Ⳛ'

SmallLetterOldShima

\u{2cdb}: 'ⳛ'

CapitalLetterOldNubianShima

\u{2cdc}: 'Ⳝ'

SmallLetterOldNubianShima

\u{2cdd}: 'ⳝ'

CapitalLetterOldNubianNgi

\u{2cde}: 'Ⳟ'

SmallLetterOldNubianNgi

\u{2cdf}: 'ⳟ'

CapitalLetterOldNubianNyi

\u{2ce0}: 'Ⳡ'

SmallLetterOldNubianNyi

\u{2ce1}: 'ⳡ'

CapitalLetterOldNubianWau

\u{2ce2}: 'Ⳣ'

SmallLetterOldNubianWau

\u{2ce3}: 'ⳣ'

SymbolKai

\u{2ce4}: 'ⳤ'

SymbolMiRo

\u{2ce5}: '⳥'

SymbolPiRo

\u{2ce6}: '⳦'

SymbolStauros

\u{2ce7}: '⳧'

SymbolTauRo

\u{2ce8}: '⳨'

SymbolKhiRo

\u{2ce9}: '⳩'

SymbolShimaSima

\u{2cea}: '⳪'

CapitalLetterCryptogrammicShei

\u{2ceb}: 'Ⳬ'

SmallLetterCryptogrammicShei

\u{2cec}: 'ⳬ'

CapitalLetterCryptogrammicGangia

\u{2ced}: 'Ⳮ'

SmallLetterCryptogrammicGangia

\u{2cee}: 'ⳮ'

CombiningNiAbove

\u{2cef}: '⳯'

CombiningSpiritusAsper

\u{2cf0}: '⳰'

CombiningSpiritusLenis

\u{2cf1}: '⳱'

CapitalLetterBohairicKhei

\u{2cf2}: 'Ⳳ'

SmallLetterBohairicKhei

\u{2cf3}: 'ⳳ'

OldNubianFullStop

\u{2cf9}: '⳹'

OldNubianDirectQuestionMark

\u{2cfa}: '⳺'

OldNubianIndirectQuestionMark

\u{2cfb}: '⳻'

OldNubianVerseDivider

\u{2cfc}: '⳼'

FractionOneHalf

\u{2cfd}: '⳽'

FullStop

\u{2cfe}: '⳾'

Methods

impl Coptic

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case