Enum unicode_types::cuneiform::Cuneiform

pub enum Cuneiform {
    SignA,
    SignATimesA,
    SignATimesBad,
    SignATimesGan2Tenu,
    SignATimesHa,
    SignATimesIgi,
    SignATimesLagarGunu,
    SignATimesMush,
    SignATimesSag,
    SignA2,
    SignAb,
    SignAbTimesAsh2,
    SignAbTimesDun3Gunu,
    SignAbTimesGal,
    SignAbTimesGan2Tenu,
    SignAbTimesHa,
    SignAbTimesIgiGunu,
    SignAbTimesImin,
    SignAbTimesLagab,
    SignAbTimesShesh,
    SignAbTimesUPlusUPlusU,
    SignAbGunu,
    SignAb2,
    SignAb2TimesBalag,
    SignAb2TimesGan2Tenu,
    SignAb2TimesMePlusEn,
    SignAb2TimesSha3,
    SignAb2TimesTak4,
    SignAd,
    SignAk,
    SignAkTimesErin2,
    SignAkTimesShitaPlusGish,
    SignAl,
    SignAlTimesAl,
    SignAlTimesDim2,
    SignAlTimesGish,
    SignAlTimesHa,
    SignAlTimesKad3,
    SignAlTimesKi,
    SignAlTimesShe,
    SignAlTimesUsh,
    SignAlan,
    SignAleph,
    SignAmar,
    SignAmarTimesShe,
    SignAn,
    SignAnOverAn,
    SignAnThreeTimes,
    SignAnPlusNagaOpposingAnPlusNaga,
    SignAnPlusNagaSquared,
    SignAnshe,
    SignApin,
    SignArad,
    SignAradTimesKur,
    SignArkab,
    SignAsal2,
    SignAsh,
    SignAshZidaTenu,
    SignAshKabaTenu,
    SignAshOverAshTug2OverTug2Tug2OverTug2Pap,
    SignAshOverAshOverAsh,
    SignAshOverAshOverAshCrossingAshOverAshOverAsh,
    SignAsh2,
    SignAshgab,
    SignBa,
    SignBad,
    SignBag3,
    SignBahar2,
    SignBal,
    SignBalOverBal,
    SignBalag,
    SignBar,
    SignBara2,
    SignBi,
    SignBiTimesA,
    SignBiTimesGar,
    SignBiTimesIgiGunu,
    SignBu,
    SignBuOverBuAb,
    SignBuOverBuUn,
    SignBuCrossingBu,
    SignBulug,
    SignBulugOverBulug,
    SignBur,
    SignBur2,
    SignDa,
    SignDag,
    SignDagKisim5TimesAPlusMash,
    SignDagKisim5TimesAmar,
    SignDagKisim5TimesBalag,
    SignDagKisim5TimesBi,
    SignDagKisim5TimesGa,
    SignDagKisim5TimesGaPlusMash,
    SignDagKisim5TimesGi,
    SignDagKisim5TimesGir2,
    SignDagKisim5TimesGud,
    SignDagKisim5TimesHa,
    SignDagKisim5TimesIr,
    SignDagKisim5TimesIrPlusLu,
    SignDagKisim5TimesKak,
    SignDagKisim5TimesLa,
    SignDagKisim5TimesLu,
    SignDagKisim5TimesLuPlusMash2,
    SignDagKisim5TimesLum,
    SignDagKisim5TimesNe,
    SignDagKisim5TimesPapPlusPap,
    SignDagKisim5TimesSi,
    SignDagKisim5TimesTak4,
    SignDagKisim5TimesU2PlusGir2,
    SignDagKisim5TimesUsh,
    SignDam,
    SignDar,
    SignDara3,
    SignDara4,
    SignDi,
    SignDib,
    SignDim,
    SignDimTimesShe,
    SignDim2,
    SignDin,
    SignDinKaskalUGunuDish,
    SignDish,
    SignDu,
    SignDuOverDu,
    SignDuGunu,
    SignDuSheshig,
    SignDub,
    SignDubTimesEsh2,
    SignDub2,
    SignDug,
    SignDugud,
    SignDuh,
    SignDun,
    SignDun3,
    SignDun3Gunu,
    SignDun3GunuGunu,
    SignDun4,
    SignDur2,
    SignE,
    SignETimesPap,
    SignEOverENunOverNun,
    SignE2,
    SignE2TimesAPlusHaPlusDa,
    SignE2TimesGar,
    SignE2TimesMi,
    SignE2TimesSal,
    SignE2TimesShe,
    SignE2TimesU,
    SignEdin,
    SignEgir,
    SignEl,
    SignEn,
    SignEnTimesGan2,
    SignEnTimesGan2Tenu,
    SignEnTimesMe,
    SignEnCrossingEn,
    SignEnOpposingEn,
    SignEnSquared,
    SignEren,
    SignErin2,
    SignEsh2,
    SignEzen,
    SignEzenTimesA,
    SignEzenTimesAPlusLal,
    SignEzenTimesAPlusLalTimesLal,
    SignEzenTimesAn,
    SignEzenTimesBad,
    SignEzenTimesDun3Gunu,
    SignEzenTimesDun3GunuGunu,
    SignEzenTimesHa,
    SignEzenTimesHaGunu,
    SignEzenTimesIgiGunu,
    SignEzenTimesKaskal,
    SignEzenTimesKaskalSquared,
    SignEzenTimesKu3,
    SignEzenTimesLa,
    SignEzenTimesLalTimesLal,
    SignEzenTimesLi,
    SignEzenTimesLu,
    SignEzenTimesU2,
    SignEzenTimesUd,
    SignGa,
    SignGaGunu,
    SignGa2,
    SignGa2TimesAPlusDaPlusHa,
    SignGa2TimesAPlusHa,
    SignGa2TimesAPlusIgi,
    SignGa2TimesAb2TenuPlusTab,
    SignGa2TimesAn,
    SignGa2TimesAsh,
    SignGa2TimesAsh2PlusGal,
    SignGa2TimesBad,
    SignGa2TimesBarPlusRa,
    SignGa2TimesBur,
    SignGa2TimesBurPlusRa,
    SignGa2TimesDa,
    SignGa2TimesDi,
    SignGa2TimesDimTimesShe,
    SignGa2TimesDub,
    SignGa2TimesEl,
    SignGa2TimesElPlusLa,
    SignGa2TimesEn,
    SignGa2TimesEnTimesGan2Tenu,
    SignGa2TimesGan2Tenu,
    SignGa2TimesGar,
    SignGa2TimesGi,
    SignGa2TimesGi4,
    SignGa2TimesGi4PlusA,
    SignGa2TimesGir2PlusSu,
    SignGa2TimesHaPlusLuPlusEsh2,
    SignGa2TimesHal,
    SignGa2TimesHalPlusLa,
    SignGa2TimesHiPlusLi,
    SignGa2TimesHub2,
    SignGa2TimesIgiGunu,
    SignGa2TimesIshPlusHuPlusAsh,
    SignGa2TimesKak,
    SignGa2TimesKaskal,
    SignGa2TimesKid,
    SignGa2TimesKidPlusLal,
    SignGa2TimesKu3PlusAn,
    SignGa2TimesLa,
    SignGa2TimesMePlusEn,
    SignGa2TimesMi,
    SignGa2TimesNun,
    SignGa2TimesNunOverNun,
    SignGa2TimesPa,
    SignGa2TimesSal,
    SignGa2TimesSar,
    SignGa2TimesShe,
    SignGa2TimesShePlusTur,
    SignGa2TimesShid,
    SignGa2TimesSum,
    SignGa2TimesTak4,
    SignGa2TimesU,
    SignGa2TimesUd,
    SignGa2TimesUdPlusDu,
    SignGa2OverGa2,
    SignGaba,
    SignGabaCrossingGaba,
    SignGad,
    SignGadOverGadGarOverGar,
    SignGal,
    SignGalGadOverGadGarOverGar,
    SignGalam,
    SignGam,
    SignGan,
    SignGan2,
    SignGan2Tenu,
    SignGan2OverGan2,
    SignGan2CrossingGan2,
    SignGar,
    SignGar3,
    SignGashan,
    SignGeshtin,
    SignGeshtinTimesKur,
    SignGi,
    SignGiTimesE,
    SignGiTimesU,
    SignGiCrossingGi,
    SignGi4,
    SignGi4OverGi4,
    SignGi4CrossingGi4,
    SignGidim,
    SignGir2,
    SignGir2Gunu,
    SignGir3,
    SignGir3TimesAPlusIgi,
    SignGir3TimesGan2Tenu,
    SignGir3TimesIgi,
    SignGir3TimesLuPlusIgi,
    SignGir3TimesPa,
    SignGisal,
    SignGish,
    SignGishCrossingGish,
    SignGishTimesBad,
    SignGishTimesTak4,
    SignGishTenu,
    SignGu,
    SignGuCrossingGu,
    SignGu2,
    SignGu2TimesKak,
    SignGu2TimesKakTimesIgiGunu,
    SignGu2TimesNun,
    SignGu2TimesSalPlusTug2,
    SignGu2Gunu,
    SignGud,
    SignGudTimesAPlusKur,
    SignGudTimesKur,
    SignGudOverGudLugal,
    SignGul,
    SignGum,
    SignGumTimesShe,
    SignGur,
    SignGur7,
    SignGurun,
    SignGurush,
    SignHa,
    SignHaTenu,
    SignHaGunu,
    SignHal,
    SignHi,
    SignHiTimesAsh,
    SignHiTimesAsh2,
    SignHiTimesBad,
    SignHiTimesDish,
    SignHiTimesGad,
    SignHiTimesKin,
    SignHiTimesNun,
    SignHiTimesShe,
    SignHiTimesU,
    SignHu,
    SignHub2,
    SignHub2TimesAn,
    SignHub2TimesHal,
    SignHub2TimesKaskal,
    SignHub2TimesLish,
    SignHub2TimesUd,
    SignHul2,
    SignI,
    SignIA,
    SignIb,
    SignIdim,
    SignIdimOverIdimBur,
    SignIdimOverIdimSquared,
    SignIg,
    SignIgi,
    SignIgiDib,
    SignIgiRi,
    SignIgiOverIgiShirOverShirUdOverUd,
    SignIgiGunu,
    SignIl,
    SignIlTimesGan2Tenu,
    SignIl2,
    SignIm,
    SignImTimesTak4,
    SignImCrossingIm,
    SignImOpposingIm,
    SignImSquared,
    SignImin,
    SignIn,
    SignIr,
    SignIsh,
    SignKa,
    SignKaTimesA,
    SignKaTimesAd,
    SignKaTimesAdPlusKu3,
    SignKaTimesAsh2,
    SignKaTimesBad,
    SignKaTimesBalag,
    SignKaTimesBar,
    SignKaTimesBi,
    SignKaTimesErin2,
    SignKaTimesEsh2,
    SignKaTimesGa,
    SignKaTimesGal,
    SignKaTimesGan2Tenu,
    SignKaTimesGar,
    SignKaTimesGarPlusSha3PlusA,
    SignKaTimesGi,
    SignKaTimesGir2,
    SignKaTimesGishPlusSar,
    SignKaTimesGishCrossingGish,
    SignKaTimesGu,
    SignKaTimesGur7,
    SignKaTimesIgi,
    SignKaTimesIm,
    SignKaTimesKak,
    SignKaTimesKi,
    SignKaTimesKid,
    SignKaTimesLi,
    SignKaTimesLu,
    SignKaTimesMe,
    SignKaTimesMePlusDu,
    SignKaTimesMePlusGi,
    SignKaTimesMePlusTe,
    SignKaTimesMi,
    SignKaTimesMiPlusNunuz,
    SignKaTimesNe,
    SignKaTimesNun,
    SignKaTimesPi,
    SignKaTimesRu,
    SignKaTimesSa,
    SignKaTimesSar,
    SignKaTimesSha,
    SignKaTimesShe,
    SignKaTimesShid,
    SignKaTimesShu,
    SignKaTimesSig,
    SignKaTimesSuhur,
    SignKaTimesTar,
    SignKaTimesU,
    SignKaTimesU2,
    SignKaTimesUd,
    SignKaTimesUmumTimesPa,
    SignKaTimesUsh,
    SignKaTimesZi,
    SignKa2,
    SignKa2CrossingKa2,
    SignKab,
    SignKad2,
    SignKad3,
    SignKad4,
    SignKad5,
    SignKad5OverKad5,
    SignKak,
    SignKakTimesIgiGunu,
    SignKal,
    SignKalTimesBad,
    SignKalCrossingKal,
    SignKam2,
    SignKam4,
    SignKaskal,
    SignKaskalLagabTimesUOverLagabTimesU,
    SignKaskalOverKaskalLagabTimesUOverLagabTimesU,
    SignKesh2,
    SignKi,
    SignKiTimesBad,
    SignKiTimesU,
    SignKiTimesUd,
    SignKid,
    SignKin,
    SignKisal,
    SignKish,
    SignKisim5,
    SignKisim5OverKisim5,
    SignKu,
    SignKuOverHiTimesAsh2KuOverHiTimesAsh2,
    SignKu3,
    SignKu4,
    SignKu4VariantForm,
    SignKu7,
    SignKul,
    SignKulGunu,
    SignKun,
    SignKur,
    SignKurOpposingKur,
    SignKushu2,
    SignKwu318,
    SignLa,
    SignLagab,
    SignLagabTimesA,
    SignLagabTimesAPlusDaPlusHa,
    SignLagabTimesAPlusGar,
    SignLagabTimesAPlusLal,
    SignLagabTimesAl,
    SignLagabTimesAn,
    SignLagabTimesAshZidaTenu,
    SignLagabTimesBad,
    SignLagabTimesBi,
    SignLagabTimesDar,
    SignLagabTimesEn,
    SignLagabTimesGa,
    SignLagabTimesGar,
    SignLagabTimesGud,
    SignLagabTimesGudPlusGud,
    SignLagabTimesHa,
    SignLagabTimesHal,
    SignLagabTimesHiTimesNun,
    SignLagabTimesIgiGunu,
    SignLagabTimesIm,
    SignLagabTimesImPlusHa,
    SignLagabTimesImPlusLu,
    SignLagabTimesKi,
    SignLagabTimesKin,
    SignLagabTimesKu3,
    SignLagabTimesKul,
    SignLagabTimesKulPlusHiPlusA,
    SignLagabTimesLagab,
    SignLagabTimesLish,
    SignLagabTimesLu,
    SignLagabTimesLul,
    SignLagabTimesMe,
    SignLagabTimesMePlusEn,
    SignLagabTimesMush,
    SignLagabTimesNe,
    SignLagabTimesShePlusSum,
    SignLagabTimesShitaPlusGishPlusErin2,
    SignLagabTimesShitaPlusGishTenu,
    SignLagabTimesShu2,
    SignLagabTimesShu2PlusShu2,
    SignLagabTimesSum,
    SignLagabTimesTag,
    SignLagabTimesTak4,
    SignLagabTimesTePlusAPlusSuPlusNa,
    SignLagabTimesU,
    SignLagabTimesUPlusA,
    SignLagabTimesUPlusUPlusU,
    SignLagabTimesU2PlusAsh,
    SignLagabTimesUd,
    SignLagabTimesUsh,
    SignLagabSquared,
    SignLagar,
    SignLagarTimesShe,
    SignLagarTimesShePlusSum,
    SignLagarGunu,
    SignLagarGunuOverLagarGunuShe,
    SignLahshu,
    SignLal,
    SignLalTimesLal,
    SignLam,
    SignLamTimesKur,
    SignLamTimesKurPlusRu,
    SignLi,
    SignLil,
    SignLimmu2,
    SignLish,
    SignLu,
    SignLuTimesBad,
    SignLu2,
    SignLu2TimesAl,
    SignLu2TimesBad,
    SignLu2TimesEsh2,
    SignLu2TimesEsh2Tenu,
    SignLu2TimesGan2Tenu,
    SignLu2TimesHiTimesBad,
    SignLu2TimesIm,
    SignLu2TimesKad2,
    SignLu2TimesKad3,
    SignLu2TimesKad3PlusAsh,
    SignLu2TimesKi,
    SignLu2TimesLaPlusAsh,
    SignLu2TimesLagab,
    SignLu2TimesMePlusEn,
    SignLu2TimesNe,
    SignLu2TimesNu,
    SignLu2TimesSiPlusAsh,
    SignLu2TimesSik2PlusBu,
    SignLu2TimesTug2,
    SignLu2Tenu,
    SignLu2CrossingLu2,
    SignLu2OpposingLu2,
    SignLu2Squared,
    SignLu2Sheshig,
    SignLu3,
    SignLugal,
    SignLugalOverLugal,
    SignLugalOpposingLugal,
    SignLugalSheshig,
    SignLuh,
    SignLul,
    SignLum,
    SignLumOverLum,
    SignLumOverLumGarOverGar,
    SignMa,
    SignMaTimesTak4,
    SignMaGunu,
    SignMa2,
    SignMah,
    SignMar,
    SignMash,
    SignMash2,
    SignMe,
    SignMes,
    SignMi,
    SignMin,
    SignMu,
    SignMuOverMu,
    SignMug,
    SignMugGunu,
    SignMunsub,
    SignMurgu2,
    SignMush,
    SignMushTimesA,
    SignMushTimesKur,
    SignMushTimesZa,
    SignMushOverMush,
    SignMushOverMushTimesAPlusNa,
    SignMushCrossingMush,
    SignMush3,
    SignMush3TimesA,
    SignMush3TimesAPlusDi,
    SignMush3TimesDi,
    SignMush3Gunu,
    SignNa,
    SignNa2,
    SignNaga,
    SignNagaInverted,
    SignNagaTimesShuTenu,
    SignNagaOpposingNaga,
    SignNagar,
    SignNamNutillu,
    SignNam,
    SignNam2,
    SignNe,
    SignNeTimesA,
    SignNeTimesUd,
    SignNeSheshig,
    SignNi,
    SignNiTimesE,
    SignNi2,
    SignNim,
    SignNimTimesGan2Tenu,
    SignNimTimesGarPlusGan2Tenu,
    SignNinda2,
    SignNinda2TimesAn,
    SignNinda2TimesAsh,
    SignNinda2TimesAshPlusAsh,
    SignNinda2TimesGud,
    SignNinda2TimesMePlusGan2Tenu,
    SignNinda2TimesNe,
    SignNinda2TimesNun,
    SignNinda2TimesShe,
    SignNinda2TimesShePlusAAn,
    SignNinda2TimesShePlusAsh,
    SignNinda2TimesShePlusAshPlusAsh,
    SignNinda2TimesU2PlusAsh,
    SignNinda2TimesUsh,
    SignNisag,
    SignNu,
    SignNu11,
    SignNun,
    SignNunLagarTimesGar,
    SignNunLagarTimesMash,
    SignNunLagarTimesSal,
    SignNunLagarTimesSalOverNunLagarTimesSal,
    SignNunLagarTimesUsh,
    SignNunTenu,
    SignNunOverNun,
    SignNunCrossingNun,
    SignNunCrossingNunLagarOverLagar,
    SignNunuz,
    SignNunuzAb2TimesAshgab,
    SignNunuzAb2TimesBi,
    SignNunuzAb2TimesDug,
    SignNunuzAb2TimesGud,
    SignNunuzAb2TimesIgiGunu,
    SignNunuzAb2TimesKad3,
    SignNunuzAb2TimesLa,
    SignNunuzAb2TimesNe,
    SignNunuzAb2TimesSila3,
    SignNunuzAb2TimesU2,
    SignNunuzKisim5TimesBi,
    SignNunuzKisim5TimesBiU,
    SignPa,
    SignPad,
    SignPan,
    SignPap,
    SignPesh2,
    SignPi,
    SignPiTimesA,
    SignPiTimesAb,
    SignPiTimesBi,
    SignPiTimesBu,
    SignPiTimesE,
    SignPiTimesI,
    SignPiTimesIb,
    SignPiTimesU,
    SignPiTimesU2,
    SignPiCrossingPi,
    SignPirig,
    SignPirigTimesKal,
    SignPirigTimesUd,
    SignPirigTimesZa,
    SignPirigOpposingPirig,
    SignRa,
    SignRab,
    SignRi,
    SignRu,
    SignSa,
    SignSagNutillu,
    SignSag,
    SignSagTimesA,
    SignSagTimesDu,
    SignSagTimesDub,
    SignSagTimesHa,
    SignSagTimesKak,
    SignSagTimesKur,
    SignSagTimesLum,
    SignSagTimesMi,
    SignSagTimesNun,
    SignSagTimesSal,
    SignSagTimesShid,
    SignSagTimesTab,
    SignSagTimesU2,
    SignSagTimesUb,
    SignSagTimesUm,
    SignSagTimesUr,
    SignSagTimesUsh,
    SignSagOverSag,
    SignSagGunu,
    SignSal,
    SignSalLagabTimesAsh2,
    SignSanga2,
    SignSar,
    SignSha,
    SignSha3,
    SignSha3TimesA,
    SignSha3TimesBad,
    SignSha3TimesGish,
    SignSha3TimesNe,
    SignSha3TimesShu2,
    SignSha3TimesTur,
    SignSha3TimesU,
    SignSha3TimesUPlusA,
    SignSha6,
    SignShab6,
    SignShar2,
    SignShe,
    SignSheHu,
    SignSheOverSheGadOverGadGarOverGar,
    SignSheOverSheTabOverTabGarOverGar,
    SignSheg9,
    SignShen,
    SignShesh,
    SignShesh2,
    SignSheshlam,
    SignShid,
    SignShidTimesA,
    SignShidTimesIm,
    SignShim,
    SignShimTimesA,
    SignShimTimesBal,
    SignShimTimesBulug,
    SignShimTimesDin,
    SignShimTimesGar,
    SignShimTimesIgi,
    SignShimTimesIgiGunu,
    SignShimTimesKushu2,
    SignShimTimesLul,
    SignShimTimesMug,
    SignShimTimesSal,
    SignShinig,
    SignShir,
    SignShirTenu,
    SignShirOverShirBurOverBur,
    SignShita,
    SignShu,
    SignShuOverInvertedShu,
    SignShu2,
    SignShubur,
    SignSi,
    SignSiGunu,
    SignSig,
    SignSig4,
    SignSig4OverSig4Shu2,
    SignSik2,
    SignSila3,
    SignSu,
    SignSuOverSu,
    SignSud,
    SignSud2,
    SignSuhur,
    SignSum,
    SignSumash,
    SignSur,
    SignSur9,
    SignTa,
    SignTaAsterisk,
    SignTaTimesHi,
    SignTaTimesMi,
    SignTaGunu,
    SignTab,
    SignTabOverTabNiOverNiDishOverDish,
    SignTabSquared,
    SignTag,
    SignTagTimesBi,
    SignTagTimesGud,
    SignTagTimesShe,
    SignTagTimesShu,
    SignTagTimesTug2,
    SignTagTimesUd,
    SignTak4,
    SignTar,
    SignTe,
    SignTeGunu,
    SignTi,
    SignTiTenu,
    SignTil,
    SignTir,
    SignTirTimesTak4,
    SignTirOverTir,
    SignTirOverTirGadOverGadGarOverGar,
    SignTu,
    SignTug2,
    SignTuk,
    SignTum,
    SignTur,
    SignTurOverTurZaOverZa,
    SignU,
    SignUGud,
    SignUUU,
    SignUOverUPaOverPaGarOverGar,
    SignUOverUSurOverSur,
    SignUOverUUReversedOverUReversed,
    SignU2,
    SignUb,
    SignUd,
    SignUdKushu2,
    SignUdTimesBad,
    SignUdTimesMi,
    SignUdTimesUPlusUPlusU,
    SignUdTimesUPlusUPlusUGunu,
    SignUdGunu,
    SignUdSheshig,
    SignUdSheshigTimesBad,
    SignUdug,
    SignUm,
    SignUmTimesLagab,
    SignUmTimesMePlusDa,
    SignUmTimesSha3,
    SignUmTimesU,
    SignUmbin,
    SignUmum,
    SignUmumTimesKaskal,
    SignUmumTimesPa,
    SignUn,
    SignUnGunu,
    SignUr,
    SignUrCrossingUr,
    SignUrSheshig,
    SignUr2,
    SignUr2TimesAPlusHa,
    SignUr2TimesAPlusNa,
    SignUr2TimesAl,
    SignUr2TimesHa,
    SignUr2TimesNun,
    SignUr2TimesU2,
    SignUr2TimesU2PlusAsh,
    SignUr2TimesU2PlusBi,
    SignUr4,
    SignUri,
    SignUri3,
    SignUru,
    SignUruTimesA,
    SignUruTimesAshgab,
    SignUruTimesBar,
    SignUruTimesDun,
    SignUruTimesGa,
    SignUruTimesGal,
    SignUruTimesGan2Tenu,
    SignUruTimesGar,
    SignUruTimesGu,
    SignUruTimesHa,
    SignUruTimesIgi,
    SignUruTimesIm,
    SignUruTimesIsh,
    SignUruTimesKi,
    SignUruTimesLum,
    SignUruTimesMin,
    SignUruTimesPa,
    SignUruTimesShe,
    SignUruTimesSig4,
    SignUruTimesTu,
    SignUruTimesUPlusGud,
    SignUruTimesUd,
    SignUruTimesUruda,
    SignUruda,
    SignUrudaTimesU,
    SignUsh,
    SignUshTimesA,
    SignUshTimesKu,
    SignUshTimesKur,
    SignUshTimesTak4,
    SignUshx,
    SignUsh2,
    SignUshumx,
    SignUtuki,
    SignUz3,
    SignUz3TimesKaskal,
    SignUzu,
    SignZa,
    SignZaTenu,
    SignZaSquaredTimesKur,
    SignZag,
    SignZamx,
    SignZe2,
    SignZi,
    SignZiOverZi,
    SignZi3,
    SignZib,
    SignZibKabaTenu,
    SignZig,
    SignZiz2,
    SignZu,
    SignZu5,
    SignZu5TimesA,
    SignZubur,
    SignZum,
    SignKapElamite,
    SignAbTimesNun,
    SignAb2TimesA,
    SignAmarTimesKug,
    SignDagKisim5TimesU2PlusMash,
    SignDag3,
    SignDishPlusShu,
    SignDubTimesShe,
    SignEzenTimesGud,
    SignEzenTimesShe,
    SignGa2TimesAnPlusKakPlusA,
    SignGa2TimesAsh2,
    SignGe22,
    SignGig,
    SignHush,
    SignKaTimesAnshe,
    SignKaTimesAsh3,
    SignKaTimesGish,
    SignKaTimesGud,
    SignKaTimesHiTimesAsh2,
    SignKaTimesLum,
    SignKaTimesPa,
    SignKaTimesShul,
    SignKaTimesTu,
    SignKaTimesUr2,
    SignLagabTimesGi,
    SignLu2SheshigTimesBad,
    SignLu2TimesEsh2PlusLal,
    SignLu2TimesShu,
    SignMesh,
    SignMush3TimesZa,
    SignNa4,
    SignNin,
    SignNin9,
    SignNinda2TimesBal,
    SignNinda2TimesGi,
    SignNu11RotatedNinetyDegrees,
    SignPesh2Asterisk,
    SignPir2,
    SignSagTimesIgiGunu,
    SignTi2,
    SignUmTimesMe,
    SignUU,
}

An enum to represent all characters in the Cuneiform block.

Variants
SignA

\u{12000}: ''

SignATimesA

\u{12001}: ''

SignATimesBad

\u{12002}: ''

SignATimesGan2Tenu

\u{12003}: ''

SignATimesHa

\u{12004}: ''

SignATimesIgi

\u{12005}: ''

SignATimesLagarGunu

\u{12006}: ''

SignATimesMush

\u{12007}: ''

SignATimesSag

\u{12008}: ''

SignA2

\u{12009}: ''

SignAb

\u{1200a}: ''

SignAbTimesAsh2

\u{1200b}: ''

SignAbTimesDun3Gunu

\u{1200c}: ''

SignAbTimesGal

\u{1200d}: ''

SignAbTimesGan2Tenu

\u{1200e}: ''

SignAbTimesHa

\u{1200f}: ''

SignAbTimesIgiGunu

\u{12010}: ''

SignAbTimesImin

\u{12011}: ''

SignAbTimesLagab

\u{12012}: ''

SignAbTimesShesh

\u{12013}: ''

SignAbTimesUPlusUPlusU

\u{12014}: ''

SignAbGunu

\u{12015}: ''

SignAb2

\u{12016}: ''

SignAb2TimesBalag

\u{12017}: ''

SignAb2TimesGan2Tenu

\u{12018}: ''

SignAb2TimesMePlusEn

\u{12019}: ''

SignAb2TimesSha3

\u{1201a}: ''

SignAb2TimesTak4

\u{1201b}: ''

SignAd

\u{1201c}: ''

SignAk

\u{1201d}: ''

SignAkTimesErin2

\u{1201e}: ''

SignAkTimesShitaPlusGish

\u{1201f}: ''

SignAl

\u{12020}: ''

SignAlTimesAl

\u{12021}: ''

SignAlTimesDim2

\u{12022}: ''

SignAlTimesGish

\u{12023}: ''

SignAlTimesHa

\u{12024}: ''

SignAlTimesKad3

\u{12025}: ''

SignAlTimesKi

\u{12026}: ''

SignAlTimesShe

\u{12027}: ''

SignAlTimesUsh

\u{12028}: ''

SignAlan

\u{12029}: ''

SignAleph

\u{1202a}: ''

SignAmar

\u{1202b}: ''

SignAmarTimesShe

\u{1202c}: ''

SignAn

\u{1202d}: ''

SignAnOverAn

\u{1202e}: ''

SignAnThreeTimes

\u{1202f}: ''

SignAnPlusNagaOpposingAnPlusNaga

\u{12030}: ''

SignAnPlusNagaSquared

\u{12031}: ''

SignAnshe

\u{12032}: ''

SignApin

\u{12033}: ''

SignArad

\u{12034}: ''

SignAradTimesKur

\u{12035}: ''

SignArkab

\u{12036}: ''

SignAsal2

\u{12037}: ''

SignAsh

\u{12038}: ''

SignAshZidaTenu

\u{12039}: ''

SignAshKabaTenu

\u{1203a}: ''

SignAshOverAshTug2OverTug2Tug2OverTug2Pap

\u{1203b}: ''

SignAshOverAshOverAsh

\u{1203c}: ''

SignAshOverAshOverAshCrossingAshOverAshOverAsh

\u{1203d}: ''

SignAsh2

\u{1203e}: ''

SignAshgab

\u{1203f}: ''

SignBa

\u{12040}: ''

SignBad

\u{12041}: ''

SignBag3

\u{12042}: ''

SignBahar2

\u{12043}: ''

SignBal

\u{12044}: ''

SignBalOverBal

\u{12045}: ''

SignBalag

\u{12046}: ''

SignBar

\u{12047}: ''

SignBara2

\u{12048}: ''

SignBi

\u{12049}: ''

SignBiTimesA

\u{1204a}: ''

SignBiTimesGar

\u{1204b}: ''

SignBiTimesIgiGunu

\u{1204c}: ''

SignBu

\u{1204d}: ''

SignBuOverBuAb

\u{1204e}: ''

SignBuOverBuUn

\u{1204f}: ''

SignBuCrossingBu

\u{12050}: ''

SignBulug

\u{12051}: ''

SignBulugOverBulug

\u{12052}: ''

SignBur

\u{12053}: ''

SignBur2

\u{12054}: ''

SignDa

\u{12055}: ''

SignDag

\u{12056}: ''

SignDagKisim5TimesAPlusMash

\u{12057}: ''

SignDagKisim5TimesAmar

\u{12058}: ''

SignDagKisim5TimesBalag

\u{12059}: ''

SignDagKisim5TimesBi

\u{1205a}: ''

SignDagKisim5TimesGa

\u{1205b}: ''

SignDagKisim5TimesGaPlusMash

\u{1205c}: ''

SignDagKisim5TimesGi

\u{1205d}: ''

SignDagKisim5TimesGir2

\u{1205e}: ''

SignDagKisim5TimesGud

\u{1205f}: ''

SignDagKisim5TimesHa

\u{12060}: ''

SignDagKisim5TimesIr

\u{12061}: ''

SignDagKisim5TimesIrPlusLu

\u{12062}: ''

SignDagKisim5TimesKak

\u{12063}: ''

SignDagKisim5TimesLa

\u{12064}: ''

SignDagKisim5TimesLu

\u{12065}: ''

SignDagKisim5TimesLuPlusMash2

\u{12066}: ''

SignDagKisim5TimesLum

\u{12067}: ''

SignDagKisim5TimesNe

\u{12068}: ''

SignDagKisim5TimesPapPlusPap

\u{12069}: ''

SignDagKisim5TimesSi

\u{1206a}: ''

SignDagKisim5TimesTak4

\u{1206b}: ''

SignDagKisim5TimesU2PlusGir2

\u{1206c}: ''

SignDagKisim5TimesUsh

\u{1206d}: ''

SignDam

\u{1206e}: ''

SignDar

\u{1206f}: ''

SignDara3

\u{12070}: ''

SignDara4

\u{12071}: ''

SignDi

\u{12072}: ''

SignDib

\u{12073}: ''

SignDim

\u{12074}: ''

SignDimTimesShe

\u{12075}: ''

SignDim2

\u{12076}: ''

SignDin

\u{12077}: ''

SignDinKaskalUGunuDish

\u{12078}: ''

SignDish

\u{12079}: ''

SignDu

\u{1207a}: ''

SignDuOverDu

\u{1207b}: ''

SignDuGunu

\u{1207c}: ''

SignDuSheshig

\u{1207d}: ''

SignDub

\u{1207e}: ''

SignDubTimesEsh2

\u{1207f}: ''

SignDub2

\u{12080}: ''

SignDug

\u{12081}: ''

SignDugud

\u{12082}: ''

SignDuh

\u{12083}: ''

SignDun

\u{12084}: ''

SignDun3

\u{12085}: ''

SignDun3Gunu

\u{12086}: ''

SignDun3GunuGunu

\u{12087}: ''

SignDun4

\u{12088}: ''

SignDur2

\u{12089}: ''

SignE

\u{1208a}: ''

SignETimesPap

\u{1208b}: ''

SignEOverENunOverNun

\u{1208c}: ''

SignE2

\u{1208d}: ''

SignE2TimesAPlusHaPlusDa

\u{1208e}: ''

SignE2TimesGar

\u{1208f}: ''

SignE2TimesMi

\u{12090}: ''

SignE2TimesSal

\u{12091}: ''

SignE2TimesShe

\u{12092}: ''

SignE2TimesU

\u{12093}: ''

SignEdin

\u{12094}: ''

SignEgir

\u{12095}: ''

SignEl

\u{12096}: ''

SignEn

\u{12097}: ''

SignEnTimesGan2

\u{12098}: ''

SignEnTimesGan2Tenu

\u{12099}: ''

SignEnTimesMe

\u{1209a}: ''

SignEnCrossingEn

\u{1209b}: ''

SignEnOpposingEn

\u{1209c}: ''

SignEnSquared

\u{1209d}: ''

SignEren

\u{1209e}: ''

SignErin2

\u{1209f}: ''

SignEsh2

\u{120a0}: ''

SignEzen

\u{120a1}: ''

SignEzenTimesA

\u{120a2}: ''

SignEzenTimesAPlusLal

\u{120a3}: ''

SignEzenTimesAPlusLalTimesLal

\u{120a4}: ''

SignEzenTimesAn

\u{120a5}: ''

SignEzenTimesBad

\u{120a6}: ''

SignEzenTimesDun3Gunu

\u{120a7}: ''

SignEzenTimesDun3GunuGunu

\u{120a8}: ''

SignEzenTimesHa

\u{120a9}: ''

SignEzenTimesHaGunu

\u{120aa}: ''

SignEzenTimesIgiGunu

\u{120ab}: ''

SignEzenTimesKaskal

\u{120ac}: ''

SignEzenTimesKaskalSquared

\u{120ad}: ''

SignEzenTimesKu3

\u{120ae}: ''

SignEzenTimesLa

\u{120af}: ''

SignEzenTimesLalTimesLal

\u{120b0}: ''

SignEzenTimesLi

\u{120b1}: ''

SignEzenTimesLu

\u{120b2}: ''

SignEzenTimesU2

\u{120b3}: ''

SignEzenTimesUd

\u{120b4}: ''

SignGa

\u{120b5}: ''

SignGaGunu

\u{120b6}: ''

SignGa2

\u{120b7}: ''

SignGa2TimesAPlusDaPlusHa

\u{120b8}: ''

SignGa2TimesAPlusHa

\u{120b9}: ''

SignGa2TimesAPlusIgi

\u{120ba}: ''

SignGa2TimesAb2TenuPlusTab

\u{120bb}: ''

SignGa2TimesAn

\u{120bc}: ''

SignGa2TimesAsh

\u{120bd}: ''

SignGa2TimesAsh2PlusGal

\u{120be}: ''

SignGa2TimesBad

\u{120bf}: ''

SignGa2TimesBarPlusRa

\u{120c0}: ''

SignGa2TimesBur

\u{120c1}: ''

SignGa2TimesBurPlusRa

\u{120c2}: ''

SignGa2TimesDa

\u{120c3}: ''

SignGa2TimesDi

\u{120c4}: ''

SignGa2TimesDimTimesShe

\u{120c5}: ''

SignGa2TimesDub

\u{120c6}: ''

SignGa2TimesEl

\u{120c7}: ''

SignGa2TimesElPlusLa

\u{120c8}: ''

SignGa2TimesEn

\u{120c9}: ''

SignGa2TimesEnTimesGan2Tenu

\u{120ca}: ''

SignGa2TimesGan2Tenu

\u{120cb}: ''

SignGa2TimesGar

\u{120cc}: ''

SignGa2TimesGi

\u{120cd}: ''

SignGa2TimesGi4

\u{120ce}: ''

SignGa2TimesGi4PlusA

\u{120cf}: ''

SignGa2TimesGir2PlusSu

\u{120d0}: ''

SignGa2TimesHaPlusLuPlusEsh2

\u{120d1}: ''

SignGa2TimesHal

\u{120d2}: ''

SignGa2TimesHalPlusLa

\u{120d3}: ''

SignGa2TimesHiPlusLi

\u{120d4}: ''

SignGa2TimesHub2

\u{120d5}: ''

SignGa2TimesIgiGunu

\u{120d6}: ''

SignGa2TimesIshPlusHuPlusAsh

\u{120d7}: ''

SignGa2TimesKak

\u{120d8}: ''

SignGa2TimesKaskal

\u{120d9}: ''

SignGa2TimesKid

\u{120da}: ''

SignGa2TimesKidPlusLal

\u{120db}: ''

SignGa2TimesKu3PlusAn

\u{120dc}: ''

SignGa2TimesLa

\u{120dd}: ''

SignGa2TimesMePlusEn

\u{120de}: ''

SignGa2TimesMi

\u{120df}: ''

SignGa2TimesNun

\u{120e0}: ''

SignGa2TimesNunOverNun

\u{120e1}: ''

SignGa2TimesPa

\u{120e2}: ''

SignGa2TimesSal

\u{120e3}: ''

SignGa2TimesSar

\u{120e4}: ''

SignGa2TimesShe

\u{120e5}: ''

SignGa2TimesShePlusTur

\u{120e6}: ''

SignGa2TimesShid

\u{120e7}: ''

SignGa2TimesSum

\u{120e8}: ''

SignGa2TimesTak4

\u{120e9}: ''

SignGa2TimesU

\u{120ea}: ''

SignGa2TimesUd

\u{120eb}: ''

SignGa2TimesUdPlusDu

\u{120ec}: ''

SignGa2OverGa2

\u{120ed}: ''

SignGaba

\u{120ee}: ''

SignGabaCrossingGaba

\u{120ef}: ''

SignGad

\u{120f0}: ''

SignGadOverGadGarOverGar

\u{120f1}: ''

SignGal

\u{120f2}: ''

SignGalGadOverGadGarOverGar

\u{120f3}: ''

SignGalam

\u{120f4}: ''

SignGam

\u{120f5}: ''

SignGan

\u{120f6}: ''

SignGan2

\u{120f7}: ''

SignGan2Tenu

\u{120f8}: ''

SignGan2OverGan2

\u{120f9}: ''

SignGan2CrossingGan2

\u{120fa}: ''

SignGar

\u{120fb}: ''

SignGar3

\u{120fc}: ''

SignGashan

\u{120fd}: ''

SignGeshtin

\u{120fe}: ''

SignGeshtinTimesKur

\u{120ff}: ''

SignGi

\u{12100}: ''

SignGiTimesE

\u{12101}: ''

SignGiTimesU

\u{12102}: ''

SignGiCrossingGi

\u{12103}: ''

SignGi4

\u{12104}: ''

SignGi4OverGi4

\u{12105}: ''

SignGi4CrossingGi4

\u{12106}: ''

SignGidim

\u{12107}: ''

SignGir2

\u{12108}: ''

SignGir2Gunu

\u{12109}: ''

SignGir3

\u{1210a}: ''

SignGir3TimesAPlusIgi

\u{1210b}: ''

SignGir3TimesGan2Tenu

\u{1210c}: ''

SignGir3TimesIgi

\u{1210d}: ''

SignGir3TimesLuPlusIgi

\u{1210e}: ''

SignGir3TimesPa

\u{1210f}: ''

SignGisal

\u{12110}: ''

SignGish

\u{12111}: ''

SignGishCrossingGish

\u{12112}: ''

SignGishTimesBad

\u{12113}: ''

SignGishTimesTak4

\u{12114}: ''

SignGishTenu

\u{12115}: ''

SignGu

\u{12116}: ''

SignGuCrossingGu

\u{12117}: ''

SignGu2

\u{12118}: ''

SignGu2TimesKak

\u{12119}: ''

SignGu2TimesKakTimesIgiGunu

\u{1211a}: ''

SignGu2TimesNun

\u{1211b}: ''

SignGu2TimesSalPlusTug2

\u{1211c}: ''

SignGu2Gunu

\u{1211d}: ''

SignGud

\u{1211e}: ''

SignGudTimesAPlusKur

\u{1211f}: ''

SignGudTimesKur

\u{12120}: ''

SignGudOverGudLugal

\u{12121}: ''

SignGul

\u{12122}: ''

SignGum

\u{12123}: ''

SignGumTimesShe

\u{12124}: ''

SignGur

\u{12125}: ''

SignGur7

\u{12126}: ''

SignGurun

\u{12127}: ''

SignGurush

\u{12128}: ''

SignHa

\u{12129}: ''

SignHaTenu

\u{1212a}: ''

SignHaGunu

\u{1212b}: ''

SignHal

\u{1212c}: ''

SignHi

\u{1212d}: ''

SignHiTimesAsh

\u{1212e}: ''

SignHiTimesAsh2

\u{1212f}: ''

SignHiTimesBad

\u{12130}: ''

SignHiTimesDish

\u{12131}: ''

SignHiTimesGad

\u{12132}: ''

SignHiTimesKin

\u{12133}: ''

SignHiTimesNun

\u{12134}: ''

SignHiTimesShe

\u{12135}: ''

SignHiTimesU

\u{12136}: ''

SignHu

\u{12137}: ''

SignHub2

\u{12138}: ''

SignHub2TimesAn

\u{12139}: ''

SignHub2TimesHal

\u{1213a}: ''

SignHub2TimesKaskal

\u{1213b}: ''

SignHub2TimesLish

\u{1213c}: ''

SignHub2TimesUd

\u{1213d}: ''

SignHul2

\u{1213e}: ''

SignI

\u{1213f}: ''

SignIA

\u{12140}: ''

SignIb

\u{12141}: ''

SignIdim

\u{12142}: ''

SignIdimOverIdimBur

\u{12143}: ''

SignIdimOverIdimSquared

\u{12144}: ''

SignIg

\u{12145}: ''

SignIgi

\u{12146}: ''

SignIgiDib

\u{12147}: ''

SignIgiRi

\u{12148}: ''

SignIgiOverIgiShirOverShirUdOverUd

\u{12149}: ''

SignIgiGunu

\u{1214a}: ''

SignIl

\u{1214b}: ''

SignIlTimesGan2Tenu

\u{1214c}: ''

SignIl2

\u{1214d}: ''

SignIm

\u{1214e}: ''

SignImTimesTak4

\u{1214f}: ''

SignImCrossingIm

\u{12150}: ''

SignImOpposingIm

\u{12151}: ''

SignImSquared

\u{12152}: ''

SignImin

\u{12153}: ''

SignIn

\u{12154}: ''

SignIr

\u{12155}: ''

SignIsh

\u{12156}: ''

SignKa

\u{12157}: ''

SignKaTimesA

\u{12158}: ''

SignKaTimesAd

\u{12159}: ''

SignKaTimesAdPlusKu3

\u{1215a}: ''

SignKaTimesAsh2

\u{1215b}: ''

SignKaTimesBad

\u{1215c}: ''

SignKaTimesBalag

\u{1215d}: ''

SignKaTimesBar

\u{1215e}: ''

SignKaTimesBi

\u{1215f}: ''

SignKaTimesErin2

\u{12160}: ''

SignKaTimesEsh2

\u{12161}: ''

SignKaTimesGa

\u{12162}: ''

SignKaTimesGal

\u{12163}: ''

SignKaTimesGan2Tenu

\u{12164}: ''

SignKaTimesGar

\u{12165}: ''

SignKaTimesGarPlusSha3PlusA

\u{12166}: ''

SignKaTimesGi

\u{12167}: ''

SignKaTimesGir2

\u{12168}: ''

SignKaTimesGishPlusSar

\u{12169}: ''

SignKaTimesGishCrossingGish

\u{1216a}: ''

SignKaTimesGu

\u{1216b}: ''

SignKaTimesGur7

\u{1216c}: ''

SignKaTimesIgi

\u{1216d}: ''

SignKaTimesIm

\u{1216e}: ''

SignKaTimesKak

\u{1216f}: ''

SignKaTimesKi

\u{12170}: ''

SignKaTimesKid

\u{12171}: ''

SignKaTimesLi

\u{12172}: ''

SignKaTimesLu

\u{12173}: ''

SignKaTimesMe

\u{12174}: ''

SignKaTimesMePlusDu

\u{12175}: ''

SignKaTimesMePlusGi

\u{12176}: ''

SignKaTimesMePlusTe

\u{12177}: ''

SignKaTimesMi

\u{12178}: ''

SignKaTimesMiPlusNunuz

\u{12179}: ''

SignKaTimesNe

\u{1217a}: ''

SignKaTimesNun

\u{1217b}: ''

SignKaTimesPi

\u{1217c}: ''

SignKaTimesRu

\u{1217d}: ''

SignKaTimesSa

\u{1217e}: ''

SignKaTimesSar

\u{1217f}: ''

SignKaTimesSha

\u{12180}: ''

SignKaTimesShe

\u{12181}: ''

SignKaTimesShid

\u{12182}: ''

SignKaTimesShu

\u{12183}: ''

SignKaTimesSig

\u{12184}: ''

SignKaTimesSuhur

\u{12185}: ''

SignKaTimesTar

\u{12186}: ''

SignKaTimesU

\u{12187}: ''

SignKaTimesU2

\u{12188}: ''

SignKaTimesUd

\u{12189}: ''

SignKaTimesUmumTimesPa

\u{1218a}: ''

SignKaTimesUsh

\u{1218b}: ''

SignKaTimesZi

\u{1218c}: ''

SignKa2

\u{1218d}: ''

SignKa2CrossingKa2

\u{1218e}: ''

SignKab

\u{1218f}: ''

SignKad2

\u{12190}: ''

SignKad3

\u{12191}: ''

SignKad4

\u{12192}: ''

SignKad5

\u{12193}: ''

SignKad5OverKad5

\u{12194}: ''

SignKak

\u{12195}: ''

SignKakTimesIgiGunu

\u{12196}: ''

SignKal

\u{12197}: ''

SignKalTimesBad

\u{12198}: ''

SignKalCrossingKal

\u{12199}: ''

SignKam2

\u{1219a}: ''

SignKam4

\u{1219b}: ''

SignKaskal

\u{1219c}: ''

SignKaskalLagabTimesUOverLagabTimesU

\u{1219d}: ''

SignKaskalOverKaskalLagabTimesUOverLagabTimesU

\u{1219e}: ''

SignKesh2

\u{1219f}: ''

SignKi

\u{121a0}: ''

SignKiTimesBad

\u{121a1}: ''

SignKiTimesU

\u{121a2}: ''

SignKiTimesUd

\u{121a3}: ''

SignKid

\u{121a4}: ''

SignKin

\u{121a5}: ''

SignKisal

\u{121a6}: ''

SignKish

\u{121a7}: ''

SignKisim5

\u{121a8}: ''

SignKisim5OverKisim5

\u{121a9}: ''

SignKu

\u{121aa}: ''

SignKuOverHiTimesAsh2KuOverHiTimesAsh2

\u{121ab}: ''

SignKu3

\u{121ac}: ''

SignKu4

\u{121ad}: ''

SignKu4VariantForm

\u{121ae}: ''

SignKu7

\u{121af}: ''

SignKul

\u{121b0}: ''

SignKulGunu

\u{121b1}: ''

SignKun

\u{121b2}: ''

SignKur

\u{121b3}: ''

SignKurOpposingKur

\u{121b4}: ''

SignKushu2

\u{121b5}: ''

SignKwu318

\u{121b6}: ''

SignLa

\u{121b7}: ''

SignLagab

\u{121b8}: ''

SignLagabTimesA

\u{121b9}: ''

SignLagabTimesAPlusDaPlusHa

\u{121ba}: ''

SignLagabTimesAPlusGar

\u{121bb}: ''

SignLagabTimesAPlusLal

\u{121bc}: ''

SignLagabTimesAl

\u{121bd}: ''

SignLagabTimesAn

\u{121be}: ''

SignLagabTimesAshZidaTenu

\u{121bf}: ''

SignLagabTimesBad

\u{121c0}: ''

SignLagabTimesBi

\u{121c1}: ''

SignLagabTimesDar

\u{121c2}: ''

SignLagabTimesEn

\u{121c3}: ''

SignLagabTimesGa

\u{121c4}: ''

SignLagabTimesGar

\u{121c5}: ''

SignLagabTimesGud

\u{121c6}: ''

SignLagabTimesGudPlusGud

\u{121c7}: ''

SignLagabTimesHa

\u{121c8}: ''

SignLagabTimesHal

\u{121c9}: ''

SignLagabTimesHiTimesNun

\u{121ca}: ''

SignLagabTimesIgiGunu

\u{121cb}: ''

SignLagabTimesIm

\u{121cc}: ''

SignLagabTimesImPlusHa

\u{121cd}: ''

SignLagabTimesImPlusLu

\u{121ce}: ''

SignLagabTimesKi

\u{121cf}: ''

SignLagabTimesKin

\u{121d0}: ''

SignLagabTimesKu3

\u{121d1}: ''

SignLagabTimesKul

\u{121d2}: ''

SignLagabTimesKulPlusHiPlusA

\u{121d3}: ''

SignLagabTimesLagab

\u{121d4}: ''

SignLagabTimesLish

\u{121d5}: ''

SignLagabTimesLu

\u{121d6}: ''

SignLagabTimesLul

\u{121d7}: ''

SignLagabTimesMe

\u{121d8}: ''

SignLagabTimesMePlusEn

\u{121d9}: ''

SignLagabTimesMush

\u{121da}: ''

SignLagabTimesNe

\u{121db}: ''

SignLagabTimesShePlusSum

\u{121dc}: ''

SignLagabTimesShitaPlusGishPlusErin2

\u{121dd}: ''

SignLagabTimesShitaPlusGishTenu

\u{121de}: ''

SignLagabTimesShu2

\u{121df}: ''

SignLagabTimesShu2PlusShu2

\u{121e0}: ''

SignLagabTimesSum

\u{121e1}: ''

SignLagabTimesTag

\u{121e2}: ''

SignLagabTimesTak4

\u{121e3}: ''

SignLagabTimesTePlusAPlusSuPlusNa

\u{121e4}: ''

SignLagabTimesU

\u{121e5}: ''

SignLagabTimesUPlusA

\u{121e6}: ''

SignLagabTimesUPlusUPlusU

\u{121e7}: ''

SignLagabTimesU2PlusAsh

\u{121e8}: ''

SignLagabTimesUd

\u{121e9}: ''

SignLagabTimesUsh

\u{121ea}: ''

SignLagabSquared

\u{121eb}: ''

SignLagar

\u{121ec}: ''

SignLagarTimesShe

\u{121ed}: ''

SignLagarTimesShePlusSum

\u{121ee}: ''

SignLagarGunu

\u{121ef}: ''

SignLagarGunuOverLagarGunuShe

\u{121f0}: ''

SignLahshu

\u{121f1}: ''

SignLal

\u{121f2}: ''

SignLalTimesLal

\u{121f3}: ''

SignLam

\u{121f4}: ''

SignLamTimesKur

\u{121f5}: ''

SignLamTimesKurPlusRu

\u{121f6}: ''

SignLi

\u{121f7}: ''

SignLil

\u{121f8}: ''

SignLimmu2

\u{121f9}: ''

SignLish

\u{121fa}: ''

SignLu

\u{121fb}: ''

SignLuTimesBad

\u{121fc}: ''

SignLu2

\u{121fd}: ''

SignLu2TimesAl

\u{121fe}: ''

SignLu2TimesBad

\u{121ff}: ''

SignLu2TimesEsh2

\u{12200}: ''

SignLu2TimesEsh2Tenu

\u{12201}: ''

SignLu2TimesGan2Tenu

\u{12202}: ''

SignLu2TimesHiTimesBad

\u{12203}: ''

SignLu2TimesIm

\u{12204}: ''

SignLu2TimesKad2

\u{12205}: ''

SignLu2TimesKad3

\u{12206}: ''

SignLu2TimesKad3PlusAsh

\u{12207}: ''

SignLu2TimesKi

\u{12208}: ''

SignLu2TimesLaPlusAsh

\u{12209}: ''

SignLu2TimesLagab

\u{1220a}: ''

SignLu2TimesMePlusEn

\u{1220b}: ''

SignLu2TimesNe

\u{1220c}: ''

SignLu2TimesNu

\u{1220d}: ''

SignLu2TimesSiPlusAsh

\u{1220e}: ''

SignLu2TimesSik2PlusBu

\u{1220f}: ''

SignLu2TimesTug2

\u{12210}: ''

SignLu2Tenu

\u{12211}: ''

SignLu2CrossingLu2

\u{12212}: ''

SignLu2OpposingLu2

\u{12213}: ''

SignLu2Squared

\u{12214}: ''

SignLu2Sheshig

\u{12215}: ''

SignLu3

\u{12216}: ''

SignLugal

\u{12217}: ''

SignLugalOverLugal

\u{12218}: ''

SignLugalOpposingLugal

\u{12219}: ''

SignLugalSheshig

\u{1221a}: ''

SignLuh

\u{1221b}: ''

SignLul

\u{1221c}: ''

SignLum

\u{1221d}: ''

SignLumOverLum

\u{1221e}: ''

SignLumOverLumGarOverGar

\u{1221f}: ''

SignMa

\u{12220}: ''

SignMaTimesTak4

\u{12221}: ''

SignMaGunu

\u{12222}: ''

SignMa2

\u{12223}: ''

SignMah

\u{12224}: ''

SignMar

\u{12225}: ''

SignMash

\u{12226}: ''

SignMash2

\u{12227}: ''

SignMe

\u{12228}: ''

SignMes

\u{12229}: ''

SignMi

\u{1222a}: ''

SignMin

\u{1222b}: ''

SignMu

\u{1222c}: ''

SignMuOverMu

\u{1222d}: ''

SignMug

\u{1222e}: ''

SignMugGunu

\u{1222f}: ''

SignMunsub

\u{12230}: ''

SignMurgu2

\u{12231}: ''

SignMush

\u{12232}: ''

SignMushTimesA

\u{12233}: ''

SignMushTimesKur

\u{12234}: ''

SignMushTimesZa

\u{12235}: ''

SignMushOverMush

\u{12236}: ''

SignMushOverMushTimesAPlusNa

\u{12237}: ''

SignMushCrossingMush

\u{12238}: ''

SignMush3

\u{12239}: ''

SignMush3TimesA

\u{1223a}: ''

SignMush3TimesAPlusDi

\u{1223b}: ''

SignMush3TimesDi

\u{1223c}: ''

SignMush3Gunu

\u{1223d}: ''

SignNa

\u{1223e}: ''

SignNa2

\u{1223f}: ''

SignNaga

\u{12240}: ''

SignNagaInverted

\u{12241}: ''

SignNagaTimesShuTenu

\u{12242}: ''

SignNagaOpposingNaga

\u{12243}: ''

SignNagar

\u{12244}: ''

SignNamNutillu

\u{12245}: ''

SignNam

\u{12246}: ''

SignNam2

\u{12247}: ''

SignNe

\u{12248}: ''

SignNeTimesA

\u{12249}: ''

SignNeTimesUd

\u{1224a}: ''

SignNeSheshig

\u{1224b}: ''

SignNi

\u{1224c}: ''

SignNiTimesE

\u{1224d}: ''

SignNi2

\u{1224e}: ''

SignNim

\u{1224f}: ''

SignNimTimesGan2Tenu

\u{12250}: ''

SignNimTimesGarPlusGan2Tenu

\u{12251}: ''

SignNinda2

\u{12252}: ''

SignNinda2TimesAn

\u{12253}: ''

SignNinda2TimesAsh

\u{12254}: ''

SignNinda2TimesAshPlusAsh

\u{12255}: ''

SignNinda2TimesGud

\u{12256}: ''

SignNinda2TimesMePlusGan2Tenu

\u{12257}: ''

SignNinda2TimesNe

\u{12258}: ''

SignNinda2TimesNun

\u{12259}: ''

SignNinda2TimesShe

\u{1225a}: ''

SignNinda2TimesShePlusAAn

\u{1225b}: ''

SignNinda2TimesShePlusAsh

\u{1225c}: ''

SignNinda2TimesShePlusAshPlusAsh

\u{1225d}: ''

SignNinda2TimesU2PlusAsh

\u{1225e}: ''

SignNinda2TimesUsh

\u{1225f}: ''

SignNisag

\u{12260}: ''

SignNu

\u{12261}: ''

SignNu11

\u{12262}: ''

SignNun

\u{12263}: ''

SignNunLagarTimesGar

\u{12264}: ''

SignNunLagarTimesMash

\u{12265}: ''

SignNunLagarTimesSal

\u{12266}: ''

SignNunLagarTimesSalOverNunLagarTimesSal

\u{12267}: ''

SignNunLagarTimesUsh

\u{12268}: ''

SignNunTenu

\u{12269}: ''

SignNunOverNun

\u{1226a}: ''

SignNunCrossingNun

\u{1226b}: ''

SignNunCrossingNunLagarOverLagar

\u{1226c}: ''

SignNunuz

\u{1226d}: ''

SignNunuzAb2TimesAshgab

\u{1226e}: ''

SignNunuzAb2TimesBi

\u{1226f}: ''

SignNunuzAb2TimesDug

\u{12270}: ''

SignNunuzAb2TimesGud

\u{12271}: ''

SignNunuzAb2TimesIgiGunu

\u{12272}: ''

SignNunuzAb2TimesKad3

\u{12273}: ''

SignNunuzAb2TimesLa

\u{12274}: ''

SignNunuzAb2TimesNe

\u{12275}: ''

SignNunuzAb2TimesSila3

\u{12276}: ''

SignNunuzAb2TimesU2

\u{12277}: ''

SignNunuzKisim5TimesBi

\u{12278}: ''

SignNunuzKisim5TimesBiU

\u{12279}: ''

SignPa

\u{1227a}: ''

SignPad

\u{1227b}: ''

SignPan

\u{1227c}: ''

SignPap

\u{1227d}: ''

SignPesh2

\u{1227e}: ''

SignPi

\u{1227f}: ''

SignPiTimesA

\u{12280}: ''

SignPiTimesAb

\u{12281}: ''

SignPiTimesBi

\u{12282}: ''

SignPiTimesBu

\u{12283}: ''

SignPiTimesE

\u{12284}: ''

SignPiTimesI

\u{12285}: ''

SignPiTimesIb

\u{12286}: ''

SignPiTimesU

\u{12287}: ''

SignPiTimesU2

\u{12288}: ''

SignPiCrossingPi

\u{12289}: ''

SignPirig

\u{1228a}: ''

SignPirigTimesKal

\u{1228b}: ''

SignPirigTimesUd

\u{1228c}: ''

SignPirigTimesZa

\u{1228d}: ''

SignPirigOpposingPirig

\u{1228e}: ''

SignRa

\u{1228f}: ''

SignRab

\u{12290}: ''

SignRi

\u{12291}: ''

SignRu

\u{12292}: ''

SignSa

\u{12293}: ''

SignSagNutillu

\u{12294}: ''

SignSag

\u{12295}: ''

SignSagTimesA

\u{12296}: ''

SignSagTimesDu

\u{12297}: ''

SignSagTimesDub

\u{12298}: ''

SignSagTimesHa

\u{12299}: ''

SignSagTimesKak

\u{1229a}: ''

SignSagTimesKur

\u{1229b}: ''

SignSagTimesLum

\u{1229c}: ''

SignSagTimesMi

\u{1229d}: ''

SignSagTimesNun

\u{1229e}: ''

SignSagTimesSal

\u{1229f}: ''

SignSagTimesShid

\u{122a0}: ''

SignSagTimesTab

\u{122a1}: ''

SignSagTimesU2

\u{122a2}: ''

SignSagTimesUb

\u{122a3}: ''

SignSagTimesUm

\u{122a4}: ''

SignSagTimesUr

\u{122a5}: ''

SignSagTimesUsh

\u{122a6}: ''

SignSagOverSag

\u{122a7}: ''

SignSagGunu

\u{122a8}: ''

SignSal

\u{122a9}: ''

SignSalLagabTimesAsh2

\u{122aa}: ''

SignSanga2

\u{122ab}: ''

SignSar

\u{122ac}: ''

SignSha

\u{122ad}: ''

SignSha3

\u{122ae}: ''

SignSha3TimesA

\u{122af}: ''

SignSha3TimesBad

\u{122b0}: ''

SignSha3TimesGish

\u{122b1}: ''

SignSha3TimesNe

\u{122b2}: ''

SignSha3TimesShu2

\u{122b3}: ''

SignSha3TimesTur

\u{122b4}: ''

SignSha3TimesU

\u{122b5}: ''

SignSha3TimesUPlusA

\u{122b6}: ''

SignSha6

\u{122b7}: ''

SignShab6

\u{122b8}: ''

SignShar2

\u{122b9}: ''

SignShe

\u{122ba}: ''

SignSheHu

\u{122bb}: ''

SignSheOverSheGadOverGadGarOverGar

\u{122bc}: ''

SignSheOverSheTabOverTabGarOverGar

\u{122bd}: ''

SignSheg9

\u{122be}: ''

SignShen

\u{122bf}: ''

SignShesh

\u{122c0}: ''

SignShesh2

\u{122c1}: ''

SignSheshlam

\u{122c2}: ''

SignShid

\u{122c3}: ''

SignShidTimesA

\u{122c4}: ''

SignShidTimesIm

\u{122c5}: ''

SignShim

\u{122c6}: ''

SignShimTimesA

\u{122c7}: ''

SignShimTimesBal

\u{122c8}: ''

SignShimTimesBulug

\u{122c9}: ''

SignShimTimesDin

\u{122ca}: ''

SignShimTimesGar

\u{122cb}: ''

SignShimTimesIgi

\u{122cc}: ''

SignShimTimesIgiGunu

\u{122cd}: ''

SignShimTimesKushu2

\u{122ce}: ''

SignShimTimesLul

\u{122cf}: ''

SignShimTimesMug

\u{122d0}: ''

SignShimTimesSal

\u{122d1}: ''

SignShinig

\u{122d2}: ''

SignShir

\u{122d3}: ''

SignShirTenu

\u{122d4}: ''

SignShirOverShirBurOverBur

\u{122d5}: ''

SignShita

\u{122d6}: ''

SignShu

\u{122d7}: ''

SignShuOverInvertedShu

\u{122d8}: ''

SignShu2

\u{122d9}: ''

SignShubur

\u{122da}: ''

SignSi

\u{122db}: ''

SignSiGunu

\u{122dc}: ''

SignSig

\u{122dd}: ''

SignSig4

\u{122de}: ''

SignSig4OverSig4Shu2

\u{122df}: ''

SignSik2

\u{122e0}: ''

SignSila3

\u{122e1}: ''

SignSu

\u{122e2}: ''

SignSuOverSu

\u{122e3}: ''

SignSud

\u{122e4}: ''

SignSud2

\u{122e5}: ''

SignSuhur

\u{122e6}: ''

SignSum

\u{122e7}: ''

SignSumash

\u{122e8}: ''

SignSur

\u{122e9}: ''

SignSur9

\u{122ea}: ''

SignTa

\u{122eb}: ''

SignTaAsterisk

\u{122ec}: ''

SignTaTimesHi

\u{122ed}: ''

SignTaTimesMi

\u{122ee}: ''

SignTaGunu

\u{122ef}: ''

SignTab

\u{122f0}: ''

SignTabOverTabNiOverNiDishOverDish

\u{122f1}: ''

SignTabSquared

\u{122f2}: ''

SignTag

\u{122f3}: ''

SignTagTimesBi

\u{122f4}: ''

SignTagTimesGud

\u{122f5}: ''

SignTagTimesShe

\u{122f6}: ''

SignTagTimesShu

\u{122f7}: ''

SignTagTimesTug2

\u{122f8}: ''

SignTagTimesUd

\u{122f9}: ''

SignTak4

\u{122fa}: ''

SignTar

\u{122fb}: ''

SignTe

\u{122fc}: ''

SignTeGunu

\u{122fd}: ''

SignTi

\u{122fe}: ''

SignTiTenu

\u{122ff}: ''

SignTil

\u{12300}: ''

SignTir

\u{12301}: ''

SignTirTimesTak4

\u{12302}: ''

SignTirOverTir

\u{12303}: ''

SignTirOverTirGadOverGadGarOverGar

\u{12304}: ''

SignTu

\u{12305}: ''

SignTug2

\u{12306}: ''

SignTuk

\u{12307}: ''

SignTum

\u{12308}: ''

SignTur

\u{12309}: ''

SignTurOverTurZaOverZa

\u{1230a}: ''

SignU

\u{1230b}: ''

SignUGud

\u{1230c}: ''

SignUUU

\u{1230d}: ''

SignUOverUPaOverPaGarOverGar

\u{1230e}: ''

SignUOverUSurOverSur

\u{1230f}: ''

SignUOverUUReversedOverUReversed

\u{12310}: ''

SignU2

\u{12311}: ''

SignUb

\u{12312}: ''

SignUd

\u{12313}: ''

SignUdKushu2

\u{12314}: ''

SignUdTimesBad

\u{12315}: ''

SignUdTimesMi

\u{12316}: ''

SignUdTimesUPlusUPlusU

\u{12317}: ''

SignUdTimesUPlusUPlusUGunu

\u{12318}: ''

SignUdGunu

\u{12319}: ''

SignUdSheshig

\u{1231a}: ''

SignUdSheshigTimesBad

\u{1231b}: ''

SignUdug

\u{1231c}: ''

SignUm

\u{1231d}: ''

SignUmTimesLagab

\u{1231e}: ''

SignUmTimesMePlusDa

\u{1231f}: ''

SignUmTimesSha3

\u{12320}: ''

SignUmTimesU

\u{12321}: ''

SignUmbin

\u{12322}: ''

SignUmum

\u{12323}: ''

SignUmumTimesKaskal

\u{12324}: ''

SignUmumTimesPa

\u{12325}: ''

SignUn

\u{12326}: ''

SignUnGunu

\u{12327}: ''

SignUr

\u{12328}: ''

SignUrCrossingUr

\u{12329}: ''

SignUrSheshig

\u{1232a}: ''

SignUr2

\u{1232b}: ''

SignUr2TimesAPlusHa

\u{1232c}: ''

SignUr2TimesAPlusNa

\u{1232d}: ''

SignUr2TimesAl

\u{1232e}: ''

SignUr2TimesHa

\u{1232f}: ''

SignUr2TimesNun

\u{12330}: ''

SignUr2TimesU2

\u{12331}: ''

SignUr2TimesU2PlusAsh

\u{12332}: ''

SignUr2TimesU2PlusBi

\u{12333}: ''

SignUr4

\u{12334}: ''

SignUri

\u{12335}: ''

SignUri3

\u{12336}: ''

SignUru

\u{12337}: ''

SignUruTimesA

\u{12338}: ''

SignUruTimesAshgab

\u{12339}: ''

SignUruTimesBar

\u{1233a}: ''

SignUruTimesDun

\u{1233b}: ''

SignUruTimesGa

\u{1233c}: ''

SignUruTimesGal

\u{1233d}: ''

SignUruTimesGan2Tenu

\u{1233e}: ''

SignUruTimesGar

\u{1233f}: ''

SignUruTimesGu

\u{12340}: ''

SignUruTimesHa

\u{12341}: ''

SignUruTimesIgi

\u{12342}: ''

SignUruTimesIm

\u{12343}: ''

SignUruTimesIsh

\u{12344}: ''

SignUruTimesKi

\u{12345}: ''

SignUruTimesLum

\u{12346}: ''

SignUruTimesMin

\u{12347}: ''

SignUruTimesPa

\u{12348}: ''

SignUruTimesShe

\u{12349}: ''

SignUruTimesSig4

\u{1234a}: ''

SignUruTimesTu

\u{1234b}: ''

SignUruTimesUPlusGud

\u{1234c}: ''

SignUruTimesUd

\u{1234d}: ''

SignUruTimesUruda

\u{1234e}: ''

SignUruda

\u{1234f}: ''

SignUrudaTimesU

\u{12350}: ''

SignUsh

\u{12351}: ''

SignUshTimesA

\u{12352}: ''

SignUshTimesKu

\u{12353}: ''

SignUshTimesKur

\u{12354}: ''

SignUshTimesTak4

\u{12355}: ''

SignUshx

\u{12356}: ''

SignUsh2

\u{12357}: ''

SignUshumx

\u{12358}: ''

SignUtuki

\u{12359}: ''

SignUz3

\u{1235a}: ''

SignUz3TimesKaskal

\u{1235b}: ''

SignUzu

\u{1235c}: ''

SignZa

\u{1235d}: ''

SignZaTenu

\u{1235e}: ''

SignZaSquaredTimesKur

\u{1235f}: ''

SignZag

\u{12360}: ''

SignZamx

\u{12361}: ''

SignZe2

\u{12362}: ''

SignZi

\u{12363}: ''

SignZiOverZi

\u{12364}: ''

SignZi3

\u{12365}: ''

SignZib

\u{12366}: ''

SignZibKabaTenu

\u{12367}: ''

SignZig

\u{12368}: ''

SignZiz2

\u{12369}: ''

SignZu

\u{1236a}: ''

SignZu5

\u{1236b}: ''

SignZu5TimesA

\u{1236c}: ''

SignZubur

\u{1236d}: ''

SignZum

\u{1236e}: ''

SignKapElamite

\u{1236f}: ''

SignAbTimesNun

\u{12370}: ''

SignAb2TimesA

\u{12371}: ''

SignAmarTimesKug

\u{12372}: ''

SignDagKisim5TimesU2PlusMash

\u{12373}: ''

SignDag3

\u{12374}: ''

SignDishPlusShu

\u{12375}: ''

SignDubTimesShe

\u{12376}: ''

SignEzenTimesGud

\u{12377}: ''

SignEzenTimesShe

\u{12378}: ''

SignGa2TimesAnPlusKakPlusA

\u{12379}: ''

SignGa2TimesAsh2

\u{1237a}: ''

SignGe22

\u{1237b}: ''

SignGig

\u{1237c}: ''

SignHush

\u{1237d}: ''

SignKaTimesAnshe

\u{1237e}: ''

SignKaTimesAsh3

\u{1237f}: ''

SignKaTimesGish

\u{12380}: ''

SignKaTimesGud

\u{12381}: ''

SignKaTimesHiTimesAsh2

\u{12382}: ''

SignKaTimesLum

\u{12383}: ''

SignKaTimesPa

\u{12384}: ''

SignKaTimesShul

\u{12385}: ''

SignKaTimesTu

\u{12386}: ''

SignKaTimesUr2

\u{12387}: ''

SignLagabTimesGi

\u{12388}: ''

SignLu2SheshigTimesBad

\u{12389}: ''

SignLu2TimesEsh2PlusLal

\u{1238a}: ''

SignLu2TimesShu

\u{1238b}: ''

SignMesh

\u{1238c}: ''

SignMush3TimesZa

\u{1238d}: ''

SignNa4

\u{1238e}: ''

SignNin

\u{1238f}: ''

SignNin9

\u{12390}: ''

SignNinda2TimesBal

\u{12391}: ''

SignNinda2TimesGi

\u{12392}: ''

SignNu11RotatedNinetyDegrees

\u{12393}: ''

SignPesh2Asterisk

\u{12394}: ''

SignPir2

\u{12395}: ''

SignSagTimesIgiGunu

\u{12396}: ''

SignTi2

\u{12397}: ''

SignUmTimesMe

\u{12398}: ''

SignUU

\u{12399}: ''

Methods

impl Cuneiform

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case