Enum unicode_types::cuneiform_numbers_and_punctuation::CuneiformNumbersandPunctuation

pub enum CuneiformNumbersandPunctuation {
    CuneiformNumericSignTwoAsh,
    CuneiformNumericSignThreeAsh,
    CuneiformNumericSignFourAsh,
    CuneiformNumericSignFiveAsh,
    CuneiformNumericSignSixAsh,
    CuneiformNumericSignSevenAsh,
    CuneiformNumericSignEightAsh,
    CuneiformNumericSignNineAsh,
    CuneiformNumericSignThreeDish,
    CuneiformNumericSignFourDish,
    CuneiformNumericSignFiveDish,
    CuneiformNumericSignSixDish,
    CuneiformNumericSignSevenDish,
    CuneiformNumericSignEightDish,
    CuneiformNumericSignNineDish,
    CuneiformNumericSignFourU,
    CuneiformNumericSignFiveU,
    CuneiformNumericSignSixU,
    CuneiformNumericSignSevenU,
    CuneiformNumericSignEightU,
    CuneiformNumericSignNineU,
    CuneiformNumericSignOneGesh2,
    CuneiformNumericSignTwoGesh2,
    CuneiformNumericSignThreeGesh2,
    CuneiformNumericSignFourGesh2,
    CuneiformNumericSignFiveGesh2,
    CuneiformNumericSignSixGesh2,
    CuneiformNumericSignSevenGesh2,
    CuneiformNumericSignEightGesh2,
    CuneiformNumericSignNineGesh2,
    CuneiformNumericSignOneGeshu,
    CuneiformNumericSignTwoGeshu,
    CuneiformNumericSignThreeGeshu,
    CuneiformNumericSignFourGeshu,
    CuneiformNumericSignFiveGeshu,
    CuneiformNumericSignTwoShar2,
    CuneiformNumericSignThreeShar2,
    CuneiformNumericSignThreeShar2VariantForm,
    CuneiformNumericSignFourShar2,
    CuneiformNumericSignFiveShar2,
    CuneiformNumericSignSixShar2,
    CuneiformNumericSignSevenShar2,
    CuneiformNumericSignEightShar2,
    CuneiformNumericSignNineShar2,
    CuneiformNumericSignOneSharu,
    CuneiformNumericSignTwoSharu,
    CuneiformNumericSignThreeSharu,
    CuneiformNumericSignThreeSharuVariantForm,
    CuneiformNumericSignFourSharu,
    CuneiformNumericSignFiveSharu,
    CuneiformNumericSignShar2TimesGalPlusDish,
    CuneiformNumericSignShar2TimesGalPlusMin,
    CuneiformNumericSignOneBuru,
    CuneiformNumericSignTwoBuru,
    CuneiformNumericSignThreeBuru,
    CuneiformNumericSignThreeBuruVariantForm,
    CuneiformNumericSignFourBuru,
    CuneiformNumericSignFiveBuru,
    CuneiformNumericSignThreeVariantFormEsh16,
    CuneiformNumericSignThreeVariantFormEsh21,
    CuneiformNumericSignFourVariantFormLimmu,
    CuneiformNumericSignFourVariantFormLimmu4,
    CuneiformNumericSignFourVariantFormLimmuA,
    CuneiformNumericSignFourVariantFormLimmuB,
    CuneiformNumericSignSixVariantFormAsh9,
    CuneiformNumericSignSevenVariantFormImin3,
    CuneiformNumericSignSevenVariantFormIminA,
    CuneiformNumericSignSevenVariantFormIminB,
    CuneiformNumericSignEightVariantFormUssu,
    CuneiformNumericSignEightVariantFormUssu3,
    CuneiformNumericSignNineVariantFormIlimmu,
    CuneiformNumericSignNineVariantFormIlimmu3,
    CuneiformNumericSignNineVariantFormIlimmu4,
    CuneiformNumericSignNineVariantFormIlimmuA,
    CuneiformNumericSignTwoAshTenu,
    CuneiformNumericSignThreeAshTenu,
    CuneiformNumericSignFourAshTenu,
    CuneiformNumericSignFiveAshTenu,
    CuneiformNumericSignSixAshTenu,
    CuneiformNumericSignOneBan2,
    CuneiformNumericSignTwoBan2,
    CuneiformNumericSignThreeBan2,
    CuneiformNumericSignFourBan2,
    CuneiformNumericSignFourBan2VariantForm,
    CuneiformNumericSignFiveBan2,
    CuneiformNumericSignFiveBan2VariantForm,
    CuneiformNumericSignNigidamin,
    CuneiformNumericSignNigidaesh,
    CuneiformNumericSignOneEshe3,
    CuneiformNumericSignTwoEshe3,
    CuneiformNumericSignOneThirdDish,
    CuneiformNumericSignTwoThirdsDish,
    CuneiformNumericSignFiveSixthsDish,
    CuneiformNumericSignOneThirdVariantFormA,
    CuneiformNumericSignTwoThirdsVariantFormA,
    CuneiformNumericSignOneEighthAsh,
    CuneiformNumericSignOneQuarterAsh,
    CuneiformNumericSignOldAssyrianOneSixth,
    CuneiformNumericSignOldAssyrianOneQuarter,
    CuneiformNumericSignOneQuarterGur,
    CuneiformNumericSignOneHalfGur,
    CuneiformNumericSignElamiteOneThird,
    CuneiformNumericSignElamiteTwoThirds,
    CuneiformNumericSignElamiteForty,
    CuneiformNumericSignElamiteFifty,
    CuneiformNumericSignFourUVariantForm,
    CuneiformNumericSignFiveUVariantForm,
    CuneiformNumericSignSixUVariantForm,
    CuneiformNumericSignSevenUVariantForm,
    CuneiformNumericSignEightUVariantForm,
    CuneiformNumericSignNineUVariantForm,
    CuneiformPunctuationSignOldAssyrianWordDivider,
    CuneiformPunctuationSignVerticalColon,
    CuneiformPunctuationSignDiagonalColon,
    CuneiformPunctuationSignDiagonalTricolon,
    CuneiformPunctuationSignDiagonalQuadcolon,
}

An enum to represent all characters in the CuneiformNumbersandPunctuation block.

Variants
CuneiformNumericSignTwoAsh

\u{12400}: 'ğ’€'

CuneiformNumericSignThreeAsh

\u{12401}: 'ğ’'

CuneiformNumericSignFourAsh

\u{12402}: 'ğ’‚'

CuneiformNumericSignFiveAsh

\u{12403}: 'ğ’ƒ'

CuneiformNumericSignSixAsh

\u{12404}: 'ğ’„'

CuneiformNumericSignSevenAsh

\u{12405}: 'ğ’…'

CuneiformNumericSignEightAsh

\u{12406}: 'ğ’†'

CuneiformNumericSignNineAsh

\u{12407}: 'ğ’‡'

CuneiformNumericSignThreeDish

\u{12408}: 'ğ’ˆ'

CuneiformNumericSignFourDish

\u{12409}: 'ğ’‰'

CuneiformNumericSignFiveDish

\u{1240a}: 'ğ’Š'

CuneiformNumericSignSixDish

\u{1240b}: 'ğ’‹'

CuneiformNumericSignSevenDish

\u{1240c}: 'ğ’Œ'

CuneiformNumericSignEightDish

\u{1240d}: 'ğ’'

CuneiformNumericSignNineDish

\u{1240e}: 'ğ’'

CuneiformNumericSignFourU

\u{1240f}: 'ğ’'

CuneiformNumericSignFiveU

\u{12410}: 'ğ’'

CuneiformNumericSignSixU

\u{12411}: 'ğ’‘'

CuneiformNumericSignSevenU

\u{12412}: 'ğ’’'

CuneiformNumericSignEightU

\u{12413}: 'ğ’“'

CuneiformNumericSignNineU

\u{12414}: 'ğ’”'

CuneiformNumericSignOneGesh2

\u{12415}: 'ğ’•'

CuneiformNumericSignTwoGesh2

\u{12416}: 'ğ’–'

CuneiformNumericSignThreeGesh2

\u{12417}: 'ğ’—'

CuneiformNumericSignFourGesh2

\u{12418}: 'ğ’˜'

CuneiformNumericSignFiveGesh2

\u{12419}: 'ğ’™'

CuneiformNumericSignSixGesh2

\u{1241a}: 'ğ’š'

CuneiformNumericSignSevenGesh2

\u{1241b}: 'ğ’›'

CuneiformNumericSignEightGesh2

\u{1241c}: 'ğ’œ'

CuneiformNumericSignNineGesh2

\u{1241d}: 'ğ’'

CuneiformNumericSignOneGeshu

\u{1241e}: 'ğ’'

CuneiformNumericSignTwoGeshu

\u{1241f}: 'ğ’Ÿ'

CuneiformNumericSignThreeGeshu

\u{12420}: 'ğ’ '

CuneiformNumericSignFourGeshu

\u{12421}: 'ğ’¡'

CuneiformNumericSignFiveGeshu

\u{12422}: 'ğ’¢'

CuneiformNumericSignTwoShar2

\u{12423}: 'ğ’£'

CuneiformNumericSignThreeShar2

\u{12424}: 'ğ’¤'

CuneiformNumericSignThreeShar2VariantForm

\u{12425}: 'ğ’¥'

CuneiformNumericSignFourShar2

\u{12426}: 'ğ’¦'

CuneiformNumericSignFiveShar2

\u{12427}: 'ğ’'

CuneiformNumericSignSixShar2

\u{12428}: 'ğ’¨'

CuneiformNumericSignSevenShar2

\u{12429}: 'ğ’©'

CuneiformNumericSignEightShar2

\u{1242a}: 'ğ’ª'

CuneiformNumericSignNineShar2

\u{1242b}: 'ğ’«'

CuneiformNumericSignOneSharu

\u{1242c}: 'ğ’¬'

CuneiformNumericSignTwoSharu

\u{1242d}: 'ğ’­'

CuneiformNumericSignThreeSharu

\u{1242e}: 'ğ’®'

CuneiformNumericSignThreeSharuVariantForm

\u{1242f}: 'ğ’¯'

CuneiformNumericSignFourSharu

\u{12430}: 'ğ’°'

CuneiformNumericSignFiveSharu

\u{12431}: 'ğ’±'

CuneiformNumericSignShar2TimesGalPlusDish

\u{12432}: 'ğ’²'

CuneiformNumericSignShar2TimesGalPlusMin

\u{12433}: 'ğ’³'

CuneiformNumericSignOneBuru

\u{12434}: 'ğ’´'

CuneiformNumericSignTwoBuru

\u{12435}: 'ğ’µ'

CuneiformNumericSignThreeBuru

\u{12436}: 'ğ’¶'

CuneiformNumericSignThreeBuruVariantForm

\u{12437}: 'ğ’·'

CuneiformNumericSignFourBuru

\u{12438}: 'ğ’¸'

CuneiformNumericSignFiveBuru

\u{12439}: 'ğ’¹'

CuneiformNumericSignThreeVariantFormEsh16

\u{1243a}: 'ğ’º'

CuneiformNumericSignThreeVariantFormEsh21

\u{1243b}: 'ğ’»'

CuneiformNumericSignFourVariantFormLimmu

\u{1243c}: 'ğ’¼'

CuneiformNumericSignFourVariantFormLimmu4

\u{1243d}: 'ğ’½'

CuneiformNumericSignFourVariantFormLimmuA

\u{1243e}: 'ğ’¾'

CuneiformNumericSignFourVariantFormLimmuB

\u{1243f}: 'ğ’¿'

CuneiformNumericSignSixVariantFormAsh9

\u{12440}: 'ğ’‘€'

CuneiformNumericSignSevenVariantFormImin3

\u{12441}: 'ğ’‘'

CuneiformNumericSignSevenVariantFormIminA

\u{12442}: 'ğ’‘‚'

CuneiformNumericSignSevenVariantFormIminB

\u{12443}: 'ğ’‘ƒ'

CuneiformNumericSignEightVariantFormUssu

\u{12444}: 'ğ’‘„'

CuneiformNumericSignEightVariantFormUssu3

\u{12445}: 'ğ’‘…'

CuneiformNumericSignNineVariantFormIlimmu

\u{12446}: 'ğ’‘†'

CuneiformNumericSignNineVariantFormIlimmu3

\u{12447}: 'ğ’‘‡'

CuneiformNumericSignNineVariantFormIlimmu4

\u{12448}: 'ğ’‘ˆ'

CuneiformNumericSignNineVariantFormIlimmuA

\u{12449}: 'ğ’‘‰'

CuneiformNumericSignTwoAshTenu

\u{1244a}: 'ğ’‘Š'

CuneiformNumericSignThreeAshTenu

\u{1244b}: 'ğ’‘‹'

CuneiformNumericSignFourAshTenu

\u{1244c}: 'ğ’‘Œ'

CuneiformNumericSignFiveAshTenu

\u{1244d}: 'ğ’‘'

CuneiformNumericSignSixAshTenu

\u{1244e}: 'ğ’‘'

CuneiformNumericSignOneBan2

\u{1244f}: 'ğ’‘'

CuneiformNumericSignTwoBan2

\u{12450}: 'ğ’‘'

CuneiformNumericSignThreeBan2

\u{12451}: 'ğ’‘‘'

CuneiformNumericSignFourBan2

\u{12452}: 'ğ’‘’'

CuneiformNumericSignFourBan2VariantForm

\u{12453}: 'ğ’‘“'

CuneiformNumericSignFiveBan2

\u{12454}: 'ğ’‘”'

CuneiformNumericSignFiveBan2VariantForm

\u{12455}: 'ğ’‘•'

CuneiformNumericSignNigidamin

\u{12456}: 'ğ’‘–'

CuneiformNumericSignNigidaesh

\u{12457}: 'ğ’‘—'

CuneiformNumericSignOneEshe3

\u{12458}: 'ğ’‘˜'

CuneiformNumericSignTwoEshe3

\u{12459}: 'ğ’‘™'

CuneiformNumericSignOneThirdDish

\u{1245a}: 'ğ’‘š'

CuneiformNumericSignTwoThirdsDish

\u{1245b}: 'ğ’‘›'

CuneiformNumericSignFiveSixthsDish

\u{1245c}: 'ğ’‘œ'

CuneiformNumericSignOneThirdVariantFormA

\u{1245d}: 'ğ’‘'

CuneiformNumericSignTwoThirdsVariantFormA

\u{1245e}: 'ğ’‘'

CuneiformNumericSignOneEighthAsh

\u{1245f}: 'ğ’‘Ÿ'

CuneiformNumericSignOneQuarterAsh

\u{12460}: 'ğ’‘ '

CuneiformNumericSignOldAssyrianOneSixth

\u{12461}: 'ğ’‘¡'

CuneiformNumericSignOldAssyrianOneQuarter

\u{12462}: 'ğ’‘¢'

CuneiformNumericSignOneQuarterGur

\u{12463}: 'ğ’‘£'

CuneiformNumericSignOneHalfGur

\u{12464}: 'ğ’‘¤'

CuneiformNumericSignElamiteOneThird

\u{12465}: 'ğ’‘¥'

CuneiformNumericSignElamiteTwoThirds

\u{12466}: 'ğ’‘¦'

CuneiformNumericSignElamiteForty

\u{12467}: 'ğ’‘'

CuneiformNumericSignElamiteFifty

\u{12468}: 'ğ’‘¨'

CuneiformNumericSignFourUVariantForm

\u{12469}: 'ğ’‘©'

CuneiformNumericSignFiveUVariantForm

\u{1246a}: 'ğ’‘ª'

CuneiformNumericSignSixUVariantForm

\u{1246b}: 'ğ’‘«'

CuneiformNumericSignSevenUVariantForm

\u{1246c}: 'ğ’‘¬'

CuneiformNumericSignEightUVariantForm

\u{1246d}: 'ğ’‘­'

CuneiformNumericSignNineUVariantForm

\u{1246e}: 'ğ’‘®'

CuneiformPunctuationSignOldAssyrianWordDivider

\u{12470}: 'ğ’‘°'

CuneiformPunctuationSignVerticalColon

\u{12471}: 'ğ’‘±'

CuneiformPunctuationSignDiagonalColon

\u{12472}: 'ğ’‘²'

CuneiformPunctuationSignDiagonalTricolon

\u{12473}: 'ğ’‘³'

CuneiformPunctuationSignDiagonalQuadcolon

\u{12474}: 'ğ’‘´'

Methods

impl CuneiformNumbersandPunctuation

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case