Enum unicode_types::cyrillic_extended_b::CyrillicExtendedB

pub enum CyrillicExtendedB {
    CyrillicCapitalLetterZemlya,
    CyrillicSmallLetterZemlya,
    CyrillicCapitalLetterDzelo,
    CyrillicSmallLetterDzelo,
    CyrillicCapitalLetterReversedDze,
    CyrillicSmallLetterReversedDze,
    CyrillicCapitalLetterIota,
    CyrillicSmallLetterIota,
    CyrillicCapitalLetterDjerv,
    CyrillicSmallLetterDjerv,
    CyrillicCapitalLetterMonographUk,
    CyrillicSmallLetterMonographUk,
    CyrillicCapitalLetterBroadOmega,
    CyrillicSmallLetterBroadOmega,
    CyrillicCapitalLetterNeutralYer,
    CyrillicSmallLetterNeutralYer,
    CyrillicCapitalLetterYeruWithBackYer,
    CyrillicSmallLetterYeruWithBackYer,
    CyrillicCapitalLetterIotifiedYat,
    CyrillicSmallLetterIotifiedYat,
    CyrillicCapitalLetterReversedYu,
    CyrillicSmallLetterReversedYu,
    CyrillicCapitalLetterIotifiedA,
    CyrillicSmallLetterIotifiedA,
    CyrillicCapitalLetterClosedLittleYus,
    CyrillicSmallLetterClosedLittleYus,
    CyrillicCapitalLetterBlendedYus,
    CyrillicSmallLetterBlendedYus,
    CyrillicCapitalLetterIotifiedClosedLittleYus,
    CyrillicSmallLetterIotifiedClosedLittleYus,
    CyrillicCapitalLetterYn,
    CyrillicSmallLetterYn,
    CyrillicCapitalLetterReversedTse,
    CyrillicSmallLetterReversedTse,
    CyrillicCapitalLetterSoftDe,
    CyrillicSmallLetterSoftDe,
    CyrillicCapitalLetterSoftEl,
    CyrillicSmallLetterSoftEl,
    CyrillicCapitalLetterSoftEm,
    CyrillicSmallLetterSoftEm,
    CyrillicCapitalLetterMonocularO,
    CyrillicSmallLetterMonocularO,
    CyrillicCapitalLetterBinocularO,
    CyrillicSmallLetterBinocularO,
    CyrillicCapitalLetterDoubleMonocularO,
    CyrillicSmallLetterDoubleMonocularO,
    CyrillicLetterMultiocularO,
    CombiningCyrillicVzmet,
    CombiningCyrillicTenMillionsSign,
    CombiningCyrillicHundredMillionsSign,
    CombiningCyrillicThousandMillionsSign,
    SlavonicAsterisk,
    CombiningCyrillicLetterUkrainianIe,
    CombiningCyrillicLetterI,
    CombiningCyrillicLetterYi,
    CombiningCyrillicLetterU,
    CombiningCyrillicLetterHardSign,
    CombiningCyrillicLetterYeru,
    CombiningCyrillicLetterSoftSign,
    CombiningCyrillicLetterOmega,
    CombiningCyrillicKavyka,
    CombiningCyrillicPayerok,
    CyrillicKavyka,
    CyrillicPayerok,
    CyrillicCapitalLetterDwe,
    CyrillicSmallLetterDwe,
    CyrillicCapitalLetterDzwe,
    CyrillicSmallLetterDzwe,
    CyrillicCapitalLetterZhwe,
    CyrillicSmallLetterZhwe,
    CyrillicCapitalLetterCche,
    CyrillicSmallLetterCche,
    CyrillicCapitalLetterDzze,
    CyrillicSmallLetterDzze,
    CyrillicCapitalLetterTeWithMiddleHook,
    CyrillicSmallLetterTeWithMiddleHook,
    CyrillicCapitalLetterTwe,
    CyrillicSmallLetterTwe,
    CyrillicCapitalLetterTswe,
    CyrillicSmallLetterTswe,
    CyrillicCapitalLetterTsse,
    CyrillicSmallLetterTsse,
    CyrillicCapitalLetterTche,
    CyrillicSmallLetterTche,
    CyrillicCapitalLetterHwe,
    CyrillicSmallLetterHwe,
    CyrillicCapitalLetterShwe,
    CyrillicSmallLetterShwe,
    CyrillicCapitalLetterDoubleO,
    CyrillicSmallLetterDoubleO,
    CyrillicCapitalLetterCrossedO,
    CyrillicSmallLetterCrossedO,
    ModifierLetterCyrillicHardSign,
    ModifierLetterCyrillicSoftSign,
    CombiningCyrillicLetterEf,
}

An enum to represent all characters in the CyrillicExtendedB block.

Variants
CyrillicCapitalLetterZemlya

\u{a640}: 'Ꙁ'

CyrillicSmallLetterZemlya

\u{a641}: 'ꙁ'

CyrillicCapitalLetterDzelo

\u{a642}: 'Ꙃ'

CyrillicSmallLetterDzelo

\u{a643}: 'ꙃ'

CyrillicCapitalLetterReversedDze

\u{a644}: 'Ꙅ'

CyrillicSmallLetterReversedDze

\u{a645}: 'ꙅ'

CyrillicCapitalLetterIota

\u{a646}: 'Ꙇ'

CyrillicSmallLetterIota

\u{a647}: 'ꙇ'

CyrillicCapitalLetterDjerv

\u{a648}: 'Ꙉ'

CyrillicSmallLetterDjerv

\u{a649}: 'ꙉ'

CyrillicCapitalLetterMonographUk

\u{a64a}: 'Ꙋ'

CyrillicSmallLetterMonographUk

\u{a64b}: 'ꙋ'

CyrillicCapitalLetterBroadOmega

\u{a64c}: 'Ꙍ'

CyrillicSmallLetterBroadOmega

\u{a64d}: 'ꙍ'

CyrillicCapitalLetterNeutralYer

\u{a64e}: 'Ꙏ'

CyrillicSmallLetterNeutralYer

\u{a64f}: 'ꙏ'

CyrillicCapitalLetterYeruWithBackYer

\u{a650}: 'Ꙑ'

CyrillicSmallLetterYeruWithBackYer

\u{a651}: 'ꙑ'

CyrillicCapitalLetterIotifiedYat

\u{a652}: 'Ꙓ'

CyrillicSmallLetterIotifiedYat

\u{a653}: 'ꙓ'

CyrillicCapitalLetterReversedYu

\u{a654}: 'Ꙕ'

CyrillicSmallLetterReversedYu

\u{a655}: 'ꙕ'

CyrillicCapitalLetterIotifiedA

\u{a656}: 'Ꙗ'

CyrillicSmallLetterIotifiedA

\u{a657}: 'ꙗ'

CyrillicCapitalLetterClosedLittleYus

\u{a658}: 'Ꙙ'

CyrillicSmallLetterClosedLittleYus

\u{a659}: 'ꙙ'

CyrillicCapitalLetterBlendedYus

\u{a65a}: 'Ꙛ'

CyrillicSmallLetterBlendedYus

\u{a65b}: 'ꙛ'

CyrillicCapitalLetterIotifiedClosedLittleYus

\u{a65c}: 'Ꙝ'

CyrillicSmallLetterIotifiedClosedLittleYus

\u{a65d}: 'ꙝ'

CyrillicCapitalLetterYn

\u{a65e}: 'Ꙟ'

CyrillicSmallLetterYn

\u{a65f}: 'ꙟ'

CyrillicCapitalLetterReversedTse

\u{a660}: 'Ꙡ'

CyrillicSmallLetterReversedTse

\u{a661}: 'ꙡ'

CyrillicCapitalLetterSoftDe

\u{a662}: 'Ꙣ'

CyrillicSmallLetterSoftDe

\u{a663}: 'ꙣ'

CyrillicCapitalLetterSoftEl

\u{a664}: 'Ꙥ'

CyrillicSmallLetterSoftEl

\u{a665}: 'ꙥ'

CyrillicCapitalLetterSoftEm

\u{a666}: 'Ꙧ'

CyrillicSmallLetterSoftEm

\u{a667}: 'ꙧ'

CyrillicCapitalLetterMonocularO

\u{a668}: 'Ꙩ'

CyrillicSmallLetterMonocularO

\u{a669}: 'ꙩ'

CyrillicCapitalLetterBinocularO

\u{a66a}: 'Ꙫ'

CyrillicSmallLetterBinocularO

\u{a66b}: 'ꙫ'

CyrillicCapitalLetterDoubleMonocularO

\u{a66c}: 'Ꙭ'

CyrillicSmallLetterDoubleMonocularO

\u{a66d}: 'ꙭ'

CyrillicLetterMultiocularO

\u{a66e}: 'ꙮ'

CombiningCyrillicVzmet

\u{a66f}: '꙯'

CombiningCyrillicTenMillionsSign

\u{a670}: '꙰'

CombiningCyrillicHundredMillionsSign

\u{a671}: '꙱'

CombiningCyrillicThousandMillionsSign

\u{a672}: '꙲'

SlavonicAsterisk

\u{a673}: '꙳'

CombiningCyrillicLetterUkrainianIe

\u{a674}: 'ꙴ'

CombiningCyrillicLetterI

\u{a675}: 'ꙵ'

CombiningCyrillicLetterYi

\u{a676}: 'ꙶ'

CombiningCyrillicLetterU

\u{a677}: 'ꙷ'

CombiningCyrillicLetterHardSign

\u{a678}: 'ꙸ'

CombiningCyrillicLetterYeru

\u{a679}: 'ꙹ'

CombiningCyrillicLetterSoftSign

\u{a67a}: 'ꙺ'

CombiningCyrillicLetterOmega

\u{a67b}: 'ꙻ'

CombiningCyrillicKavyka

\u{a67c}: '꙼'

CombiningCyrillicPayerok

\u{a67d}: '꙽'

CyrillicKavyka

\u{a67e}: '꙾'

CyrillicPayerok

\u{a67f}: 'ꙿ'

CyrillicCapitalLetterDwe

\u{a680}: 'Ꚁ'

CyrillicSmallLetterDwe

\u{a681}: 'ꚁ'

CyrillicCapitalLetterDzwe

\u{a682}: 'Ꚃ'

CyrillicSmallLetterDzwe

\u{a683}: 'ꚃ'

CyrillicCapitalLetterZhwe

\u{a684}: 'Ꚅ'

CyrillicSmallLetterZhwe

\u{a685}: 'ꚅ'

CyrillicCapitalLetterCche

\u{a686}: 'Ꚇ'

CyrillicSmallLetterCche

\u{a687}: 'ꚇ'

CyrillicCapitalLetterDzze

\u{a688}: 'Ꚉ'

CyrillicSmallLetterDzze

\u{a689}: 'ꚉ'

CyrillicCapitalLetterTeWithMiddleHook

\u{a68a}: 'Ꚋ'

CyrillicSmallLetterTeWithMiddleHook

\u{a68b}: 'ꚋ'

CyrillicCapitalLetterTwe

\u{a68c}: 'Ꚍ'

CyrillicSmallLetterTwe

\u{a68d}: 'ꚍ'

CyrillicCapitalLetterTswe

\u{a68e}: 'Ꚏ'

CyrillicSmallLetterTswe

\u{a68f}: 'ꚏ'

CyrillicCapitalLetterTsse

\u{a690}: 'Ꚑ'

CyrillicSmallLetterTsse

\u{a691}: 'ꚑ'

CyrillicCapitalLetterTche

\u{a692}: 'Ꚓ'

CyrillicSmallLetterTche

\u{a693}: 'ꚓ'

CyrillicCapitalLetterHwe

\u{a694}: 'Ꚕ'

CyrillicSmallLetterHwe

\u{a695}: 'ꚕ'

CyrillicCapitalLetterShwe

\u{a696}: 'Ꚗ'

CyrillicSmallLetterShwe

\u{a697}: 'ꚗ'

CyrillicCapitalLetterDoubleO

\u{a698}: 'Ꚙ'

CyrillicSmallLetterDoubleO

\u{a699}: 'ꚙ'

CyrillicCapitalLetterCrossedO

\u{a69a}: 'Ꚛ'

CyrillicSmallLetterCrossedO

\u{a69b}: 'ꚛ'

ModifierLetterCyrillicHardSign

\u{a69c}: 'ꚜ'

ModifierLetterCyrillicSoftSign

\u{a69d}: 'ꚝ'

CombiningCyrillicLetterEf

\u{a69e}: 'ꚞ'

Methods

impl CyrillicExtendedB

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case