Enum unicode_types::deseret::Deseret

pub enum Deseret {
    CapitalLetterLongI,
    CapitalLetterLongE,
    CapitalLetterLongA,
    CapitalLetterLongAh,
    CapitalLetterLongO,
    CapitalLetterLongOo,
    CapitalLetterShortI,
    CapitalLetterShortE,
    CapitalLetterShortA,
    CapitalLetterShortAh,
    CapitalLetterShortO,
    CapitalLetterShortOo,
    CapitalLetterAy,
    CapitalLetterOw,
    CapitalLetterWu,
    CapitalLetterYee,
    CapitalLetterH,
    CapitalLetterPee,
    CapitalLetterBee,
    CapitalLetterTee,
    CapitalLetterDee,
    CapitalLetterChee,
    CapitalLetterJee,
    CapitalLetterKay,
    CapitalLetterGay,
    CapitalLetterEf,
    CapitalLetterVee,
    CapitalLetterEth,
    CapitalLetterThee,
    CapitalLetterEs,
    CapitalLetterZee,
    CapitalLetterEsh,
    CapitalLetterZhee,
    CapitalLetterEr,
    CapitalLetterEl,
    CapitalLetterEm,
    CapitalLetterEn,
    CapitalLetterEng,
    CapitalLetterOi,
    CapitalLetterEw,
    SmallLetterLongI,
    SmallLetterLongE,
    SmallLetterLongA,
    SmallLetterLongAh,
    SmallLetterLongO,
    SmallLetterLongOo,
    SmallLetterShortI,
    SmallLetterShortE,
    SmallLetterShortA,
    SmallLetterShortAh,
    SmallLetterShortO,
    SmallLetterShortOo,
    SmallLetterAy,
    SmallLetterOw,
    SmallLetterWu,
    SmallLetterYee,
    SmallLetterH,
    SmallLetterPee,
    SmallLetterBee,
    SmallLetterTee,
    SmallLetterDee,
    SmallLetterChee,
    SmallLetterJee,
    SmallLetterKay,
    SmallLetterGay,
    SmallLetterEf,
    SmallLetterVee,
    SmallLetterEth,
    SmallLetterThee,
    SmallLetterEs,
    SmallLetterZee,
    SmallLetterEsh,
    SmallLetterZhee,
    SmallLetterEr,
    SmallLetterEl,
    SmallLetterEm,
    SmallLetterEn,
    SmallLetterEng,
    SmallLetterOi,
}

An enum to represent all characters in the Deseret block.

Variants
CapitalLetterLongI

\u{10400}: 'ğ€'

CapitalLetterLongE

\u{10401}: 'ğ'

CapitalLetterLongA

\u{10402}: 'ğ‚'

CapitalLetterLongAh

\u{10403}: 'ğƒ'

CapitalLetterLongO

\u{10404}: 'ğ„'

CapitalLetterLongOo

\u{10405}: 'ğ…'

CapitalLetterShortI

\u{10406}: 'ğ†'

CapitalLetterShortE

\u{10407}: 'ğ‡'

CapitalLetterShortA

\u{10408}: 'ğˆ'

CapitalLetterShortAh

\u{10409}: 'ğ‰'

CapitalLetterShortO

\u{1040a}: 'ğŠ'

CapitalLetterShortOo

\u{1040b}: 'ğ‹'

CapitalLetterAy

\u{1040c}: 'ğŒ'

CapitalLetterOw

\u{1040d}: 'ğ'

CapitalLetterWu

\u{1040e}: 'ğ'

CapitalLetterYee

\u{1040f}: 'ğ'

CapitalLetterH

\u{10410}: 'ğ'

CapitalLetterPee

\u{10411}: 'ğ‘'

CapitalLetterBee

\u{10412}: 'ğ’'

CapitalLetterTee

\u{10413}: 'ğ“'

CapitalLetterDee

\u{10414}: 'ğ”'

CapitalLetterChee

\u{10415}: 'ğ•'

CapitalLetterJee

\u{10416}: 'ğ–'

CapitalLetterKay

\u{10417}: 'ğ—'

CapitalLetterGay

\u{10418}: 'ğ˜'

CapitalLetterEf

\u{10419}: 'ğ™'

CapitalLetterVee

\u{1041a}: 'ğš'

CapitalLetterEth

\u{1041b}: 'ğ›'

CapitalLetterThee

\u{1041c}: 'ğœ'

CapitalLetterEs

\u{1041d}: 'ğ'

CapitalLetterZee

\u{1041e}: 'ğ'

CapitalLetterEsh

\u{1041f}: 'ğŸ'

CapitalLetterZhee

\u{10420}: 'ğ '

CapitalLetterEr

\u{10421}: 'ğ¡'

CapitalLetterEl

\u{10422}: 'ğ¢'

CapitalLetterEm

\u{10423}: 'ğ£'

CapitalLetterEn

\u{10424}: 'ğ¤'

CapitalLetterEng

\u{10425}: 'ğ¥'

CapitalLetterOi

\u{10426}: 'ğ¦'

CapitalLetterEw

\u{10427}: 'ğ'

SmallLetterLongI

\u{10428}: 'ğ¨'

SmallLetterLongE

\u{10429}: 'ğ©'

SmallLetterLongA

\u{1042a}: 'ğª'

SmallLetterLongAh

\u{1042b}: 'ğ«'

SmallLetterLongO

\u{1042c}: 'ğ¬'

SmallLetterLongOo

\u{1042d}: 'ğ­'

SmallLetterShortI

\u{1042e}: 'ğ®'

SmallLetterShortE

\u{1042f}: 'ğ¯'

SmallLetterShortA

\u{10430}: 'ğ°'

SmallLetterShortAh

\u{10431}: 'ğ±'

SmallLetterShortO

\u{10432}: 'ğ²'

SmallLetterShortOo

\u{10433}: 'ğ³'

SmallLetterAy

\u{10434}: 'ğ´'

SmallLetterOw

\u{10435}: 'ğµ'

SmallLetterWu

\u{10436}: 'ğ¶'

SmallLetterYee

\u{10437}: 'ğ·'

SmallLetterH

\u{10438}: 'ğ¸'

SmallLetterPee

\u{10439}: 'ğ¹'

SmallLetterBee

\u{1043a}: 'ğº'

SmallLetterTee

\u{1043b}: 'ğ»'

SmallLetterDee

\u{1043c}: 'ğ¼'

SmallLetterChee

\u{1043d}: 'ğ½'

SmallLetterJee

\u{1043e}: 'ğ¾'

SmallLetterKay

\u{1043f}: 'ğ¿'

SmallLetterGay

\u{10440}: 'ğ‘€'

SmallLetterEf

\u{10441}: 'ğ‘'

SmallLetterVee

\u{10442}: 'ğ‘‚'

SmallLetterEth

\u{10443}: 'ğ‘ƒ'

SmallLetterThee

\u{10444}: 'ğ‘„'

SmallLetterEs

\u{10445}: 'ğ‘…'

SmallLetterZee

\u{10446}: 'ğ‘†'

SmallLetterEsh

\u{10447}: 'ğ‘‡'

SmallLetterZhee

\u{10448}: 'ğ‘ˆ'

SmallLetterEr

\u{10449}: 'ğ‘‰'

SmallLetterEl

\u{1044a}: 'ğ‘Š'

SmallLetterEm

\u{1044b}: 'ğ‘‹'

SmallLetterEn

\u{1044c}: 'ğ‘Œ'

SmallLetterEng

\u{1044d}: 'ğ‘'

SmallLetterOi

\u{1044e}: 'ğ‘'

Methods

impl Deseret

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case