Enum unicode_types::dingbats::Dingbats

pub enum Dingbats {
    BlackSafetyScissors,
    UpperBladeScissors,
    BlackScissors,
    LowerBladeScissors,
    WhiteScissors,
    WhiteHeavyCheckMark,
    TelephoneLocationSign,
    TapeDrive,
    Airplane,
    Envelope,
    RaisedFist,
    RaisedHand,
    VictoryHand,
    WritingHand,
    LowerRightPencil,
    Pencil,
    UpperRightPencil,
    WhiteNib,
    BlackNib,
    CheckMark,
    HeavyCheckMark,
    MultiplicationX,
    HeavyMultiplicationX,
    BallotX,
    HeavyBallotX,
    OutlinedGreekCross,
    HeavyGreekCross,
    OpenCentreCross,
    HeavyOpenCentreCross,
    LatinCross,
    ShadowedWhiteLatinCross,
    OutlinedLatinCross,
    MalteseCross,
    StarOfDavid,
    FourTeardropDashSpokedAsterisk,
    FourBalloonDashSpokedAsterisk,
    HeavyFourBalloonDashSpokedAsterisk,
    FourClubDashSpokedAsterisk,
    BlackFourPointedStar,
    WhiteFourPointedStar,
    Sparkles,
    StressOutlinedWhiteStar,
    CircledWhiteStar,
    OpenCentreBlackStar,
    BlackCentreWhiteStar,
    OutlinedBlackStar,
    HeavyOutlinedBlackStar,
    PinwheelStar,
    ShadowedWhiteStar,
    HeavyAsterisk,
    OpenCentreAsterisk,
    EightSpokedAsterisk,
    EightPointedBlackStar,
    EightPointedPinwheelStar,
    SixPointedBlackStar,
    EightPointedRectilinearBlackStar,
    HeavyEightPointedRectilinearBlackStar,
    TwelvePointedBlackStar,
    SixteenPointedAsterisk,
    TeardropDashSpokedAsterisk,
    OpenCentreTeardropDashSpokedAsterisk,
    HeavyTeardropDashSpokedAsterisk,
    SixPetalledBlackAndWhiteFlorette,
    BlackFlorette,
    WhiteFlorette,
    EightPetalledOutlinedBlackFlorette,
    CircledOpenCentreEightPointedStar,
    HeavyTeardropDashSpokedPinwheelAsterisk,
    Snowflake,
    TightTrifoliateSnowflake,
    HeavyChevronSnowflake,
    Sparkle,
    HeavySparkle,
    BalloonDashSpokedAsterisk,
    EightTeardropDashSpokedPropellerAsterisk,
    HeavyEightTeardropDashSpokedPropellerAsterisk,
    CrossMark,
    ShadowedWhiteCircle,
    NegativeSquaredCrossMark,
    LowerRightDropDashShadowedWhiteSquare,
    UpperRightDropDashShadowedWhiteSquare,
    LowerRightShadowedWhiteSquare,
    UpperRightShadowedWhiteSquare,
    BlackQuestionMarkOrnament,
    WhiteQuestionMarkOrnament,
    WhiteExclamationMarkOrnament,
    BlackDiamondMinusWhiteX,
    HeavyExclamationMarkSymbol,
    LightVerticalBar,
    MediumVerticalBar,
    HeavyVerticalBar,
    HeavySingleTurnedCommaQuotationMarkOrnament,
    HeavySingleCommaQuotationMarkOrnament,
    HeavyDoubleTurnedCommaQuotationMarkOrnament,
    HeavyDoubleCommaQuotationMarkOrnament,
    HeavyLowSingleCommaQuotationMarkOrnament,
    HeavyLowDoubleCommaQuotationMarkOrnament,
    CurvedStemParagraphSignOrnament,
    HeavyExclamationMarkOrnament,
    HeavyHeartExclamationMarkOrnament,
    HeavyBlackHeart,
    RotatedHeavyBlackHeartBullet,
    FloralHeart,
    RotatedFloralHeartBullet,
    MediumLeftParenthesisOrnament,
    MediumRightParenthesisOrnament,
    MediumFlattenedLeftParenthesisOrnament,
    MediumFlattenedRightParenthesisOrnament,
    MediumLeftDashPointingAngleBracketOrnament,
    MediumRightDashPointingAngleBracketOrnament,
    HeavyLeftDashPointingAngleQuotationMarkOrnament,
    HeavyRightDashPointingAngleQuotationMarkOrnament,
    HeavyLeftDashPointingAngleBracketOrnament,
    HeavyRightDashPointingAngleBracketOrnament,
    LightLeftTortoiseShellBracketOrnament,
    LightRightTortoiseShellBracketOrnament,
    MediumLeftCurlyBracketOrnament,
    MediumRightCurlyBracketOrnament,
    DingbatNegativeCircledDigitOne,
    DingbatNegativeCircledDigitTwo,
    DingbatNegativeCircledDigitThree,
    DingbatNegativeCircledDigitFour,
    DingbatNegativeCircledDigitFive,
    DingbatNegativeCircledDigitSix,
    DingbatNegativeCircledDigitSeven,
    DingbatNegativeCircledDigitEight,
    DingbatNegativeCircledDigitNine,
    DingbatNegativeCircledNumberTen,
    DingbatCircledSansDashSerifDigitOne,
    DingbatCircledSansDashSerifDigitTwo,
    DingbatCircledSansDashSerifDigitThree,
    DingbatCircledSansDashSerifDigitFour,
    DingbatCircledSansDashSerifDigitFive,
    DingbatCircledSansDashSerifDigitSix,
    DingbatCircledSansDashSerifDigitSeven,
    DingbatCircledSansDashSerifDigitEight,
    DingbatCircledSansDashSerifDigitNine,
    DingbatCircledSansDashSerifNumberTen,
    DingbatNegativeCircledSansDashSerifDigitOne,
    DingbatNegativeCircledSansDashSerifDigitTwo,
    DingbatNegativeCircledSansDashSerifDigitThree,
    DingbatNegativeCircledSansDashSerifDigitFour,
    DingbatNegativeCircledSansDashSerifDigitFive,
    DingbatNegativeCircledSansDashSerifDigitSix,
    DingbatNegativeCircledSansDashSerifDigitSeven,
    DingbatNegativeCircledSansDashSerifDigitEight,
    DingbatNegativeCircledSansDashSerifDigitNine,
    DingbatNegativeCircledSansDashSerifNumberTen,
    HeavyWideDashHeadedRightwardsArrow,
    HeavyPlusSign,
    HeavyMinusSign,
    HeavyDivisionSign,
    HeavySouthEastArrow,
    HeavyRightwardsArrow,
    HeavyNorthEastArrow,
    DraftingPointRightwardsArrow,
    HeavyRoundDashTippedRightwardsArrow,
    TriangleDashHeadedRightwardsArrow,
    HeavyTriangleDashHeadedRightwardsArrow,
    DashedTriangleDashHeadedRightwardsArrow,
    HeavyDashedTriangleDashHeadedRightwardsArrow,
    BlackRightwardsArrow,
    ThreeDashDTopDashLightedRightwardsArrowhead,
    ThreeDashDBottomDashLightedRightwardsArrowhead,
    BlackRightwardsArrowhead,
    HeavyBlackCurvedDownwardsAndRightwardsArrow,
    HeavyBlackCurvedUpwardsAndRightwardsArrow,
    SquatBlackRightwardsArrow,
    HeavyConcaveDashPointedBlackRightwardsArrow,
    RightDashShadedWhiteRightwardsArrow,
    LeftDashShadedWhiteRightwardsArrow,
    BackDashTiltedShadowedWhiteRightwardsArrow,
    FrontDashTiltedShadowedWhiteRightwardsArrow,
    HeavyLowerRightDashShadowedWhiteRightwardsArrow,
    HeavyUpperRightDashShadowedWhiteRightwardsArrow,
    NotchedLowerRightDashShadowedWhiteRightwardsArrow,
    CurlyLoop,
    NotchedUpperRightDashShadowedWhiteRightwardsArrow,
    CircledHeavyWhiteRightwardsArrow,
    WhiteDashFeatheredRightwardsArrow,
    BlackDashFeatheredSouthEastArrow,
    BlackDashFeatheredRightwardsArrow,
    BlackDashFeatheredNorthEastArrow,
    HeavyBlackDashFeatheredSouthEastArrow,
    HeavyBlackDashFeatheredRightwardsArrow,
    HeavyBlackDashFeatheredNorthEastArrow,
    TeardropDashBarbedRightwardsArrow,
    HeavyTeardropDashShankedRightwardsArrow,
    WedgeDashTailedRightwardsArrow,
    HeavyWedgeDashTailedRightwardsArrow,
    OpenDashOutlinedRightwardsArrow,
}

An enum to represent all characters in the Dingbats block.

Variants
BlackSafetyScissors

\u{2700}: '✀'

UpperBladeScissors

\u{2701}: '✁'

BlackScissors

\u{2702}: '✂'

LowerBladeScissors

\u{2703}: '✃'

WhiteScissors

\u{2704}: '✄'

WhiteHeavyCheckMark

\u{2705}: '✅'

TelephoneLocationSign

\u{2706}: '✆'

TapeDrive

\u{2707}: '✇'

Airplane

\u{2708}: '✈'

Envelope

\u{2709}: '✉'

RaisedFist

\u{270a}: '✊'

RaisedHand

\u{270b}: '✋'

VictoryHand

\u{270c}: '✌'

WritingHand

\u{270d}: '✍'

LowerRightPencil

\u{270e}: '✎'

Pencil

\u{270f}: '✏'

UpperRightPencil

\u{2710}: '✐'

WhiteNib

\u{2711}: '✑'

BlackNib

\u{2712}: '✒'

CheckMark

\u{2713}: '✓'

HeavyCheckMark

\u{2714}: '✔'

MultiplicationX

\u{2715}: '✕'

HeavyMultiplicationX

\u{2716}: '✖'

BallotX

\u{2717}: '✗'

HeavyBallotX

\u{2718}: '✘'

OutlinedGreekCross

\u{2719}: '✙'

HeavyGreekCross

\u{271a}: '✚'

OpenCentreCross

\u{271b}: '✛'

HeavyOpenCentreCross

\u{271c}: '✜'

LatinCross

\u{271d}: '✝'

ShadowedWhiteLatinCross

\u{271e}: '✞'

OutlinedLatinCross

\u{271f}: '✟'

MalteseCross

\u{2720}: '✠'

StarOfDavid

\u{2721}: '✡'

FourTeardropDashSpokedAsterisk

\u{2722}: '✢'

FourBalloonDashSpokedAsterisk

\u{2723}: '✣'

HeavyFourBalloonDashSpokedAsterisk

\u{2724}: '✤'

FourClubDashSpokedAsterisk

\u{2725}: '✥'

BlackFourPointedStar

\u{2726}: '✦'

WhiteFourPointedStar

\u{2727}: '✧'

Sparkles

\u{2728}: '✨'

StressOutlinedWhiteStar

\u{2729}: '✩'

CircledWhiteStar

\u{272a}: '✪'

OpenCentreBlackStar

\u{272b}: '✫'

BlackCentreWhiteStar

\u{272c}: '✬'

OutlinedBlackStar

\u{272d}: '✭'

HeavyOutlinedBlackStar

\u{272e}: '✮'

PinwheelStar

\u{272f}: '✯'

ShadowedWhiteStar

\u{2730}: '✰'

HeavyAsterisk

\u{2731}: '✱'

OpenCentreAsterisk

\u{2732}: '✲'

EightSpokedAsterisk

\u{2733}: '✳'

EightPointedBlackStar

\u{2734}: '✴'

EightPointedPinwheelStar

\u{2735}: '✵'

SixPointedBlackStar

\u{2736}: '✶'

EightPointedRectilinearBlackStar

\u{2737}: '✷'

HeavyEightPointedRectilinearBlackStar

\u{2738}: '✸'

TwelvePointedBlackStar

\u{2739}: '✹'

SixteenPointedAsterisk

\u{273a}: '✺'

TeardropDashSpokedAsterisk

\u{273b}: '✻'

OpenCentreTeardropDashSpokedAsterisk

\u{273c}: '✼'

HeavyTeardropDashSpokedAsterisk

\u{273d}: '✽'

SixPetalledBlackAndWhiteFlorette

\u{273e}: '✾'

BlackFlorette

\u{273f}: '✿'

WhiteFlorette

\u{2740}: '❀'

EightPetalledOutlinedBlackFlorette

\u{2741}: '❁'

CircledOpenCentreEightPointedStar

\u{2742}: '❂'

HeavyTeardropDashSpokedPinwheelAsterisk

\u{2743}: '❃'

Snowflake

\u{2744}: '❄'

TightTrifoliateSnowflake

\u{2745}: '❅'

HeavyChevronSnowflake

\u{2746}: '❆'

Sparkle

\u{2747}: '❇'

HeavySparkle

\u{2748}: '❈'

BalloonDashSpokedAsterisk

\u{2749}: '❉'

EightTeardropDashSpokedPropellerAsterisk

\u{274a}: '❊'

HeavyEightTeardropDashSpokedPropellerAsterisk

\u{274b}: '❋'

CrossMark

\u{274c}: '❌'

ShadowedWhiteCircle

\u{274d}: '❍'

NegativeSquaredCrossMark

\u{274e}: '❎'

LowerRightDropDashShadowedWhiteSquare

\u{274f}: '❏'

UpperRightDropDashShadowedWhiteSquare

\u{2750}: '❐'

LowerRightShadowedWhiteSquare

\u{2751}: '❑'

UpperRightShadowedWhiteSquare

\u{2752}: '❒'

BlackQuestionMarkOrnament

\u{2753}: '❓'

WhiteQuestionMarkOrnament

\u{2754}: '❔'

WhiteExclamationMarkOrnament

\u{2755}: '❕'

BlackDiamondMinusWhiteX

\u{2756}: '❖'

HeavyExclamationMarkSymbol

\u{2757}: '❗'

LightVerticalBar

\u{2758}: '❘'

MediumVerticalBar

\u{2759}: '❙'

HeavyVerticalBar

\u{275a}: '❚'

HeavySingleTurnedCommaQuotationMarkOrnament

\u{275b}: '❛'

HeavySingleCommaQuotationMarkOrnament

\u{275c}: '❜'

HeavyDoubleTurnedCommaQuotationMarkOrnament

\u{275d}: '❝'

HeavyDoubleCommaQuotationMarkOrnament

\u{275e}: '❞'

HeavyLowSingleCommaQuotationMarkOrnament

\u{275f}: '❟'

HeavyLowDoubleCommaQuotationMarkOrnament

\u{2760}: '❠'

CurvedStemParagraphSignOrnament

\u{2761}: '❡'

HeavyExclamationMarkOrnament

\u{2762}: '❢'

HeavyHeartExclamationMarkOrnament

\u{2763}: '❣'

HeavyBlackHeart

\u{2764}: '❤'

RotatedHeavyBlackHeartBullet

\u{2765}: '❥'

FloralHeart

\u{2766}: '❦'

RotatedFloralHeartBullet

\u{2767}: '❧'

MediumLeftParenthesisOrnament

\u{2768}: '❨'

MediumRightParenthesisOrnament

\u{2769}: '❩'

MediumFlattenedLeftParenthesisOrnament

\u{276a}: '❪'

MediumFlattenedRightParenthesisOrnament

\u{276b}: '❫'

MediumLeftDashPointingAngleBracketOrnament

\u{276c}: '❬'

MediumRightDashPointingAngleBracketOrnament

\u{276d}: '❭'

HeavyLeftDashPointingAngleQuotationMarkOrnament

\u{276e}: '❮'

HeavyRightDashPointingAngleQuotationMarkOrnament

\u{276f}: '❯'

HeavyLeftDashPointingAngleBracketOrnament

\u{2770}: '❰'

HeavyRightDashPointingAngleBracketOrnament

\u{2771}: '❱'

LightLeftTortoiseShellBracketOrnament

\u{2772}: '❲'

LightRightTortoiseShellBracketOrnament

\u{2773}: '❳'

MediumLeftCurlyBracketOrnament

\u{2774}: '❴'

MediumRightCurlyBracketOrnament

\u{2775}: '❵'

DingbatNegativeCircledDigitOne

\u{2776}: '❶'

DingbatNegativeCircledDigitTwo

\u{2777}: '❷'

DingbatNegativeCircledDigitThree

\u{2778}: '❸'

DingbatNegativeCircledDigitFour

\u{2779}: '❹'

DingbatNegativeCircledDigitFive

\u{277a}: '❺'

DingbatNegativeCircledDigitSix

\u{277b}: '❻'

DingbatNegativeCircledDigitSeven

\u{277c}: '❼'

DingbatNegativeCircledDigitEight

\u{277d}: '❽'

DingbatNegativeCircledDigitNine

\u{277e}: '❾'

DingbatNegativeCircledNumberTen

\u{277f}: '❿'

DingbatCircledSansDashSerifDigitOne

\u{2780}: '➀'

DingbatCircledSansDashSerifDigitTwo

\u{2781}: '➁'

DingbatCircledSansDashSerifDigitThree

\u{2782}: '➂'

DingbatCircledSansDashSerifDigitFour

\u{2783}: '➃'

DingbatCircledSansDashSerifDigitFive

\u{2784}: '➄'

DingbatCircledSansDashSerifDigitSix

\u{2785}: '➅'

DingbatCircledSansDashSerifDigitSeven

\u{2786}: '➆'

DingbatCircledSansDashSerifDigitEight

\u{2787}: '➇'

DingbatCircledSansDashSerifDigitNine

\u{2788}: '➈'

DingbatCircledSansDashSerifNumberTen

\u{2789}: '➉'

DingbatNegativeCircledSansDashSerifDigitOne

\u{278a}: '➊'

DingbatNegativeCircledSansDashSerifDigitTwo

\u{278b}: '➋'

DingbatNegativeCircledSansDashSerifDigitThree

\u{278c}: '➌'

DingbatNegativeCircledSansDashSerifDigitFour

\u{278d}: '➍'

DingbatNegativeCircledSansDashSerifDigitFive

\u{278e}: '➎'

DingbatNegativeCircledSansDashSerifDigitSix

\u{278f}: '➏'

DingbatNegativeCircledSansDashSerifDigitSeven

\u{2790}: '➐'

DingbatNegativeCircledSansDashSerifDigitEight

\u{2791}: '➑'

DingbatNegativeCircledSansDashSerifDigitNine

\u{2792}: '➒'

DingbatNegativeCircledSansDashSerifNumberTen

\u{2793}: '➓'

HeavyWideDashHeadedRightwardsArrow

\u{2794}: '➔'

HeavyPlusSign

\u{2795}: '➕'

HeavyMinusSign

\u{2796}: '➖'

HeavyDivisionSign

\u{2797}: '➗'

HeavySouthEastArrow

\u{2798}: '➘'

HeavyRightwardsArrow

\u{2799}: '➙'

HeavyNorthEastArrow

\u{279a}: '➚'

DraftingPointRightwardsArrow

\u{279b}: '➛'

HeavyRoundDashTippedRightwardsArrow

\u{279c}: '➜'

TriangleDashHeadedRightwardsArrow

\u{279d}: '➝'

HeavyTriangleDashHeadedRightwardsArrow

\u{279e}: '➞'

DashedTriangleDashHeadedRightwardsArrow

\u{279f}: '➟'

HeavyDashedTriangleDashHeadedRightwardsArrow

\u{27a0}: '➠'

BlackRightwardsArrow

\u{27a1}: '➡'

ThreeDashDTopDashLightedRightwardsArrowhead

\u{27a2}: '➢'

ThreeDashDBottomDashLightedRightwardsArrowhead

\u{27a3}: '➣'

BlackRightwardsArrowhead

\u{27a4}: '➤'

HeavyBlackCurvedDownwardsAndRightwardsArrow

\u{27a5}: '➥'

HeavyBlackCurvedUpwardsAndRightwardsArrow

\u{27a6}: '➦'

SquatBlackRightwardsArrow

\u{27a7}: '➧'

HeavyConcaveDashPointedBlackRightwardsArrow

\u{27a8}: '➨'

RightDashShadedWhiteRightwardsArrow

\u{27a9}: '➩'

LeftDashShadedWhiteRightwardsArrow

\u{27aa}: '➪'

BackDashTiltedShadowedWhiteRightwardsArrow

\u{27ab}: '➫'

FrontDashTiltedShadowedWhiteRightwardsArrow

\u{27ac}: '➬'

HeavyLowerRightDashShadowedWhiteRightwardsArrow

\u{27ad}: '➭'

HeavyUpperRightDashShadowedWhiteRightwardsArrow

\u{27ae}: '➮'

NotchedLowerRightDashShadowedWhiteRightwardsArrow

\u{27af}: '➯'

CurlyLoop

\u{27b0}: '➰'

NotchedUpperRightDashShadowedWhiteRightwardsArrow

\u{27b1}: '➱'

CircledHeavyWhiteRightwardsArrow

\u{27b2}: '➲'

WhiteDashFeatheredRightwardsArrow

\u{27b3}: '➳'

BlackDashFeatheredSouthEastArrow

\u{27b4}: '➴'

BlackDashFeatheredRightwardsArrow

\u{27b5}: '➵'

BlackDashFeatheredNorthEastArrow

\u{27b6}: '➶'

HeavyBlackDashFeatheredSouthEastArrow

\u{27b7}: '➷'

HeavyBlackDashFeatheredRightwardsArrow

\u{27b8}: '➸'

HeavyBlackDashFeatheredNorthEastArrow

\u{27b9}: '➹'

TeardropDashBarbedRightwardsArrow

\u{27ba}: '➺'

HeavyTeardropDashShankedRightwardsArrow

\u{27bb}: '➻'

WedgeDashTailedRightwardsArrow

\u{27bc}: '➼'

HeavyWedgeDashTailedRightwardsArrow

\u{27bd}: '➽'

OpenDashOutlinedRightwardsArrow

\u{27be}: '➾'

Methods

impl Dingbats

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case