Enum unicode_types::duployan::Duployan

pub enum Duployan {
    LetterH,
    LetterX,
    LetterP,
    LetterT,
    LetterF,
    LetterK,
    LetterL,
    LetterB,
    LetterD,
    LetterV,
    LetterG,
    LetterR,
    LetterPN,
    LetterDS,
    LetterFN,
    LetterKM,
    LetterRS,
    LetterTh,
    LetterSloanDh,
    LetterDh,
    LetterKk,
    LetterSloanJ,
    LetterHl,
    LetterLh,
    LetterRh,
    LetterM,
    LetterN,
    LetterJ,
    LetterS,
    LetterMN,
    LetterNM,
    LetterJM,
    LetterSJ,
    LetterMWithDot,
    LetterNWithDot,
    LetterJWithDot,
    LetterJWithDotsInsideAndAbove,
    LetterSWithDot,
    LetterSWithDotBelow,
    LetterMS,
    LetterNS,
    LetterJS,
    LetterSS,
    LetterMNS,
    LetterNMS,
    LetterJMS,
    LetterSJS,
    LetterJSWithDot,
    LetterJN,
    LetterJNS,
    LetterST,
    LetterSTR,
    LetterSP,
    LetterSPR,
    LetterTS,
    LetterTRS,
    LetterW,
    LetterWh,
    LetterWR,
    LetterSN,
    LetterSM,
    LetterKRS,
    LetterGRS,
    LetterSK,
    LetterSKR,
    LetterA,
    LetterSloanOw,
    LetterOa,
    LetterO,
    LetterAou,
    LetterI,
    LetterE,
    LetterIe,
    LetterShortI,
    LetterUi,
    LetterEe,
    LetterSloanEh,
    LetterRomanianI,
    LetterSloanEe,
    LetterLongI,
    LetterYe,
    LetterU,
    LetterEu,
    LetterXw,
    LetterUN,
    LetterLongU,
    LetterRomanianU,
    LetterUh,
    LetterSloanU,
    LetterOoh,
    LetterOw,
    LetterOu,
    LetterWa,
    LetterWo,
    LetterWi,
    LetterWei,
    LetterWow,
    LetterNasalU,
    LetterNasalO,
    LetterNasalI,
    LetterNasalA,
    LetterPerninAn,
    LetterPerninAm,
    LetterSloanEn,
    LetterSloanAn,
    LetterSloanOn,
    LetterVocalicM,
    AffixLeftHorizontalSecant,
    AffixMidHorizontalSecant,
    AffixRightHorizontalSecant,
    AffixLowVerticalSecant,
    AffixMidVerticalSecant,
    AffixHighVerticalSecant,
    AffixAttachedSecant,
    AffixAttachedLeftDashToDashRightSecant,
    AffixAttachedTangent,
    AffixAttachedTail,
    AffixAttachedEHook,
    AffixAttachedIHook,
    AffixAttachedTangentHook,
    AffixHighAcute,
    AffixHighTightAcute,
    AffixHighGrave,
    AffixHighLongGrave,
    AffixHighDot,
    AffixHighCircle,
    AffixHighLine,
    AffixHighWave,
    AffixHighVertical,
    AffixLowAcute,
    AffixLowTightAcute,
    AffixLowGrave,
    AffixLowLongGrave,
    AffixLowDot,
    AffixLowCircle,
    AffixLowLine,
    AffixLowWave,
    AffixLowVertical,
    AffixLowArrow,
    SignOWithCross,
    ThickLetterSelector,
    DoubleMark,
}

An enum to represent all characters in the Duployan block.

Variants
LetterH

\u{1bc00}: 'ğ›°€'

LetterX

\u{1bc01}: 'ğ›°'

LetterP

\u{1bc02}: 'ğ›°‚'

LetterT

\u{1bc03}: 'ğ›°ƒ'

LetterF

\u{1bc04}: 'ğ›°„'

LetterK

\u{1bc05}: 'ğ›°…'

LetterL

\u{1bc06}: 'ğ›°†'

LetterB

\u{1bc07}: 'ğ›°‡'

LetterD

\u{1bc08}: 'ğ›°ˆ'

LetterV

\u{1bc09}: 'ğ›°‰'

LetterG

\u{1bc0a}: 'ğ›°Š'

LetterR

\u{1bc0b}: 'ğ›°‹'

LetterPN

\u{1bc0c}: 'ğ›°Œ'

LetterDS

\u{1bc0d}: 'ğ›°'

LetterFN

\u{1bc0e}: 'ğ›°'

LetterKM

\u{1bc0f}: 'ğ›°'

LetterRS

\u{1bc10}: 'ğ›°'

LetterTh

\u{1bc11}: 'ğ›°‘'

LetterSloanDh

\u{1bc12}: 'ğ›°’'

LetterDh

\u{1bc13}: 'ğ›°“'

LetterKk

\u{1bc14}: 'ğ›°”'

LetterSloanJ

\u{1bc15}: 'ğ›°•'

LetterHl

\u{1bc16}: 'ğ›°–'

LetterLh

\u{1bc17}: 'ğ›°—'

LetterRh

\u{1bc18}: 'ğ›°˜'

LetterM

\u{1bc19}: 'ğ›°™'

LetterN

\u{1bc1a}: 'ğ›°š'

LetterJ

\u{1bc1b}: 'ğ›°›'

LetterS

\u{1bc1c}: 'ğ›°œ'

LetterMN

\u{1bc1d}: 'ğ›°'

LetterNM

\u{1bc1e}: 'ğ›°'

LetterJM

\u{1bc1f}: 'ğ›°Ÿ'

LetterSJ

\u{1bc20}: 'ğ›° '

LetterMWithDot

\u{1bc21}: 'ğ›°¡'

LetterNWithDot

\u{1bc22}: 'ğ›°¢'

LetterJWithDot

\u{1bc23}: 'ğ›°£'

LetterJWithDotsInsideAndAbove

\u{1bc24}: 'ğ›°¤'

LetterSWithDot

\u{1bc25}: 'ğ›°¥'

LetterSWithDotBelow

\u{1bc26}: 'ğ›°¦'

LetterMS

\u{1bc27}: 'ğ›°'

LetterNS

\u{1bc28}: 'ğ›°¨'

LetterJS

\u{1bc29}: 'ğ›°©'

LetterSS

\u{1bc2a}: 'ğ›°ª'

LetterMNS

\u{1bc2b}: 'ğ›°«'

LetterNMS

\u{1bc2c}: 'ğ›°¬'

LetterJMS

\u{1bc2d}: 'ğ›°­'

LetterSJS

\u{1bc2e}: 'ğ›°®'

LetterJSWithDot

\u{1bc2f}: 'ğ›°¯'

LetterJN

\u{1bc30}: 'ğ›°°'

LetterJNS

\u{1bc31}: 'ğ›°±'

LetterST

\u{1bc32}: 'ğ›°²'

LetterSTR

\u{1bc33}: 'ğ›°³'

LetterSP

\u{1bc34}: 'ğ›°´'

LetterSPR

\u{1bc35}: 'ğ›°µ'

LetterTS

\u{1bc36}: 'ğ›°¶'

LetterTRS

\u{1bc37}: 'ğ›°·'

LetterW

\u{1bc38}: 'ğ›°¸'

LetterWh

\u{1bc39}: 'ğ›°¹'

LetterWR

\u{1bc3a}: 'ğ›°º'

LetterSN

\u{1bc3b}: 'ğ›°»'

LetterSM

\u{1bc3c}: 'ğ›°¼'

LetterKRS

\u{1bc3d}: 'ğ›°½'

LetterGRS

\u{1bc3e}: 'ğ›°¾'

LetterSK

\u{1bc3f}: 'ğ›°¿'

LetterSKR

\u{1bc40}: 'ğ›±€'

LetterA

\u{1bc41}: 'ğ›±'

LetterSloanOw

\u{1bc42}: 'ğ›±‚'

LetterOa

\u{1bc43}: 'ğ›±ƒ'

LetterO

\u{1bc44}: 'ğ›±„'

LetterAou

\u{1bc45}: 'ğ›±…'

LetterI

\u{1bc46}: 'ğ›±†'

LetterE

\u{1bc47}: 'ğ›±‡'

LetterIe

\u{1bc48}: 'ğ›±ˆ'

LetterShortI

\u{1bc49}: 'ğ›±‰'

LetterUi

\u{1bc4a}: 'ğ›±Š'

LetterEe

\u{1bc4b}: 'ğ›±‹'

LetterSloanEh

\u{1bc4c}: 'ğ›±Œ'

LetterRomanianI

\u{1bc4d}: 'ğ›±'

LetterSloanEe

\u{1bc4e}: 'ğ›±'

LetterLongI

\u{1bc4f}: 'ğ›±'

LetterYe

\u{1bc50}: 'ğ›±'

LetterU

\u{1bc51}: 'ğ›±‘'

LetterEu

\u{1bc52}: 'ğ›±’'

LetterXw

\u{1bc53}: 'ğ›±“'

LetterUN

\u{1bc54}: 'ğ›±”'

LetterLongU

\u{1bc55}: 'ğ›±•'

LetterRomanianU

\u{1bc56}: 'ğ›±–'

LetterUh

\u{1bc57}: 'ğ›±—'

LetterSloanU

\u{1bc58}: 'ğ›±˜'

LetterOoh

\u{1bc59}: 'ğ›±™'

LetterOw

\u{1bc5a}: 'ğ›±š'

LetterOu

\u{1bc5b}: 'ğ›±›'

LetterWa

\u{1bc5c}: 'ğ›±œ'

LetterWo

\u{1bc5d}: 'ğ›±'

LetterWi

\u{1bc5e}: 'ğ›±'

LetterWei

\u{1bc5f}: 'ğ›±Ÿ'

LetterWow

\u{1bc60}: 'ğ›± '

LetterNasalU

\u{1bc61}: 'ğ›±¡'

LetterNasalO

\u{1bc62}: 'ğ›±¢'

LetterNasalI

\u{1bc63}: 'ğ›±£'

LetterNasalA

\u{1bc64}: 'ğ›±¤'

LetterPerninAn

\u{1bc65}: 'ğ›±¥'

LetterPerninAm

\u{1bc66}: 'ğ›±¦'

LetterSloanEn

\u{1bc67}: 'ğ›±'

LetterSloanAn

\u{1bc68}: 'ğ›±¨'

LetterSloanOn

\u{1bc69}: 'ğ›±©'

LetterVocalicM

\u{1bc6a}: 'ğ›±ª'

AffixLeftHorizontalSecant

\u{1bc70}: 'ğ›±°'

AffixMidHorizontalSecant

\u{1bc71}: 'ğ›±±'

AffixRightHorizontalSecant

\u{1bc72}: 'ğ›±²'

AffixLowVerticalSecant

\u{1bc73}: 'ğ›±³'

AffixMidVerticalSecant

\u{1bc74}: 'ğ›±´'

AffixHighVerticalSecant

\u{1bc75}: 'ğ›±µ'

AffixAttachedSecant

\u{1bc76}: 'ğ›±¶'

AffixAttachedLeftDashToDashRightSecant

\u{1bc77}: 'ğ›±·'

AffixAttachedTangent

\u{1bc78}: 'ğ›±¸'

AffixAttachedTail

\u{1bc79}: 'ğ›±¹'

AffixAttachedEHook

\u{1bc7a}: 'ğ›±º'

AffixAttachedIHook

\u{1bc7b}: 'ğ›±»'

AffixAttachedTangentHook

\u{1bc7c}: 'ğ›±¼'

AffixHighAcute

\u{1bc80}: 'ğ›²€'

AffixHighTightAcute

\u{1bc81}: 'ğ›²'

AffixHighGrave

\u{1bc82}: 'ğ›²‚'

AffixHighLongGrave

\u{1bc83}: 'ğ›²ƒ'

AffixHighDot

\u{1bc84}: 'ğ›²„'

AffixHighCircle

\u{1bc85}: 'ğ›²…'

AffixHighLine

\u{1bc86}: 'ğ›²†'

AffixHighWave

\u{1bc87}: 'ğ›²‡'

AffixHighVertical

\u{1bc88}: 'ğ›²ˆ'

AffixLowAcute

\u{1bc90}: 'ğ›²'

AffixLowTightAcute

\u{1bc91}: 'ğ›²‘'

AffixLowGrave

\u{1bc92}: 'ğ›²’'

AffixLowLongGrave

\u{1bc93}: 'ğ›²“'

AffixLowDot

\u{1bc94}: 'ğ›²”'

AffixLowCircle

\u{1bc95}: 'ğ›²•'

AffixLowLine

\u{1bc96}: 'ğ›²–'

AffixLowWave

\u{1bc97}: 'ğ›²—'

AffixLowVertical

\u{1bc98}: 'ğ›²˜'

AffixLowArrow

\u{1bc99}: 'ğ›²™'

SignOWithCross

\u{1bc9c}: 'ğ›²œ'

ThickLetterSelector

\u{1bc9d}: 'ğ›²'

DoubleMark

\u{1bc9e}: 'ğ›²'

Methods

impl Duployan

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case