Enum unicode_types::early_dynastic_cuneiform::EarlyDynasticCuneiform

pub enum EarlyDynasticCuneiform {
    CuneiformSignAbTimesNunTenu,
    CuneiformSignAbTimesShu2,
    CuneiformSignAdTimesEsh2,
    CuneiformSignBadTimesDishTenu,
    CuneiformSignBahar2TimesAb2,
    CuneiformSignBahar2TimesNi,
    CuneiformSignBahar2TimesZa,
    CuneiformSignBuOverBuTimesNa2,
    CuneiformSignDaTimesTak4,
    CuneiformSignDagTimesKur,
    CuneiformSignDimTimesIgi,
    CuneiformSignDimTimesUUU,
    CuneiformSignDim2TimesUd,
    CuneiformSignDugTimesAnshe,
    CuneiformSignDugTimesAsh,
    CuneiformSignDugTimesAshAtLeft,
    CuneiformSignDugTimesDin,
    CuneiformSignDugTimesDun,
    CuneiformSignDugTimesErin2,
    CuneiformSignDugTimesGa,
    CuneiformSignDugTimesGi,
    CuneiformSignDugTimesGir2Gunu,
    CuneiformSignDugTimesGish,
    CuneiformSignDugTimesHa,
    CuneiformSignDugTimesHi,
    CuneiformSignDugTimesIgiGunu,
    CuneiformSignDugTimesKaskal,
    CuneiformSignDugTimesKur,
    CuneiformSignDugTimesKushu2,
    CuneiformSignDugTimesKushu2PlusKaskal,
    CuneiformSignDugTimesLakDash020,
    CuneiformSignDugTimesLam,
    CuneiformSignDugTimesLamTimesKur,
    CuneiformSignDugTimesLuhPlusGish,
    CuneiformSignDugTimesMash,
    CuneiformSignDugTimesMes,
    CuneiformSignDugTimesMi,
    CuneiformSignDugTimesNi,
    CuneiformSignDugTimesPi,
    CuneiformSignDugTimesShe,
    CuneiformSignDugTimesSiGunu,
    CuneiformSignE2TimesKur,
    CuneiformSignE2TimesPap,
    CuneiformSignErin2X,
    CuneiformSignEsh2CrossingEsh2,
    CuneiformSignEzenSheshigTimesAsh,
    CuneiformSignEzenSheshigTimesHi,
    CuneiformSignEzenSheshigTimesIgiGunu,
    CuneiformSignEzenSheshigTimesLa,
    CuneiformSignEzenSheshigTimesLal,
    CuneiformSignEzenSheshigTimesMe,
    CuneiformSignEzenSheshigTimesMes,
    CuneiformSignEzenSheshigTimesSu,
    CuneiformSignEzenTimesSu,
    CuneiformSignGa2TimesBahar2,
    CuneiformSignGa2TimesDimGunu,
    CuneiformSignGa2TimesDugTimesIgiGunu,
    CuneiformSignGa2TimesDugTimesKaskal,
    CuneiformSignGa2TimesEren,
    CuneiformSignGa2TimesGa,
    CuneiformSignGa2TimesGarPlusDi,
    CuneiformSignGa2TimesGarPlusNe,
    CuneiformSignGa2TimesHaPlusA,
    CuneiformSignGa2TimesKushu2PlusKaskal,
    CuneiformSignGa2TimesLam,
    CuneiformSignGa2TimesLamTimesKur,
    CuneiformSignGa2TimesLuh,
    CuneiformSignGa2TimesMush,
    CuneiformSignGa2TimesNe,
    CuneiformSignGa2TimesNePlusE2,
    CuneiformSignGa2TimesNePlusGi,
    CuneiformSignGa2TimesShim,
    CuneiformSignGa2TimesZiz2,
    CuneiformSignGabaRotatedNinetyDegrees,
    CuneiformSignGeshtinTimesU,
    CuneiformSignGishTimesGishCrossingGish,
    CuneiformSignGu2TimesIgiGunu,
    CuneiformSignGudPlusGishTimesTak4,
    CuneiformSignHaTenuGunu,
    CuneiformSignHiTimesAshOverHiTimesAsh,
    CuneiformSignKaTimesBu,
    CuneiformSignKaTimesKa,
    CuneiformSignKaTimesUUU,
    CuneiformSignKaTimesUr,
    CuneiformSignLagabTimesZuOverZu,
    CuneiformSignLakDash003,
    CuneiformSignLakDash021,
    CuneiformSignLakDash025,
    CuneiformSignLakDash030,
    CuneiformSignLakDash050,
    CuneiformSignLakDash051,
    CuneiformSignLakDash062,
    CuneiformSignLakDash079OverLakDash079Gunu,
    CuneiformSignLakDash080,
    CuneiformSignLakDash081OverLakDash081,
    CuneiformSignLakDash092,
    CuneiformSignLakDash130,
    CuneiformSignLakDash142,
    CuneiformSignLakDash210,
    CuneiformSignLakDash219,
    CuneiformSignLakDash220,
    CuneiformSignLakDash225,
    CuneiformSignLakDash228,
    CuneiformSignLakDash238,
    CuneiformSignLakDash265,
    CuneiformSignLakDash266,
    CuneiformSignLakDash343,
    CuneiformSignLakDash347,
    CuneiformSignLakDash348,
    CuneiformSignLakDash383,
    CuneiformSignLakDash384,
    CuneiformSignLakDash390,
    CuneiformSignLakDash441,
    CuneiformSignLakDash449,
    CuneiformSignLakDash449TimesGu,
    CuneiformSignLakDash449TimesIgi,
    CuneiformSignLakDash449TimesPapPlusLu3,
    CuneiformSignLakDash449TimesPapPlusPapPlusLu3,
    CuneiformSignLakDash449TimesU2PlusBa,
    CuneiformSignLakDash450,
    CuneiformSignLakDash457,
    CuneiformSignLakDash470,
    CuneiformSignLakDash483,
    CuneiformSignLakDash490,
    CuneiformSignLakDash492,
    CuneiformSignLakDash493,
    CuneiformSignLakDash495,
    CuneiformSignLakDash550,
    CuneiformSignLakDash608,
    CuneiformSignLakDash617,
    CuneiformSignLakDash617TimesAsh,
    CuneiformSignLakDash617TimesBad,
    CuneiformSignLakDash617TimesDun3GunuGunu,
    CuneiformSignLakDash617TimesKu3,
    CuneiformSignLakDash617TimesLa,
    CuneiformSignLakDash617TimesTar,
    CuneiformSignLakDash617TimesTe,
    CuneiformSignLakDash617TimesU2,
    CuneiformSignLakDash617TimesUd,
    CuneiformSignLakDash617TimesUruda,
    CuneiformSignLakDash636,
    CuneiformSignLakDash648,
    CuneiformSignLakDash648TimesDub,
    CuneiformSignLakDash648TimesGa,
    CuneiformSignLakDash648TimesIgi,
    CuneiformSignLakDash648TimesIgiGunu,
    CuneiformSignLakDash648TimesNi,
    CuneiformSignLakDash648TimesPapPlusPapPlusLu3,
    CuneiformSignLakDash648TimesSheshPlusKi,
    CuneiformSignLakDash648TimesUd,
    CuneiformSignLakDash648TimesUruda,
    CuneiformSignLakDash724,
    CuneiformSignLakDash749,
    CuneiformSignLu2GunuTimesAsh,
    CuneiformSignLu2TimesDish,
    CuneiformSignLu2TimesHal,
    CuneiformSignLu2TimesPap,
    CuneiformSignLu2TimesPapPlusPapPlusLu3,
    CuneiformSignLu2TimesTak4,
    CuneiformSignMiPlusZa7,
    CuneiformSignMushOverMushTimesGa,
    CuneiformSignMushOverMushTimesKak,
    CuneiformSignNinda2TimesDimGunu,
    CuneiformSignNinda2TimesGish,
    CuneiformSignNinda2TimesGul,
    CuneiformSignNinda2TimesHi,
    CuneiformSignNinda2TimesKesh2,
    CuneiformSignNinda2TimesLakDash050,
    CuneiformSignNinda2TimesMash,
    CuneiformSignNinda2TimesPapPlusPap,
    CuneiformSignNinda2TimesU,
    CuneiformSignNinda2TimesUPlusU,
    CuneiformSignNinda2TimesUruda,
    CuneiformSignSagGunuTimesHa,
    CuneiformSignSagTimesEn,
    CuneiformSignSagTimesSheAtLeft,
    CuneiformSignSagTimesTak4,
    CuneiformSignSha6Tenu,
    CuneiformSignSheOverShe,
    CuneiformSignShePlusHub2,
    CuneiformSignShePlusNam2,
    CuneiformSignShePlusSar,
    CuneiformSignShu2PlusDugTimesNi,
    CuneiformSignShu2PlusE2TimesAn,
    CuneiformSignSiTimesTak4,
    CuneiformSignTak4PlusSag,
    CuneiformSignTumTimesGan2Tenu,
    CuneiformSignTumTimesThreeDish,
    CuneiformSignUr2Inverted,
    CuneiformSignUr2TimesUd,
    CuneiformSignUruTimesDara3,
    CuneiformSignUruTimesLakDash668,
    CuneiformSignUruTimesLu3,
    CuneiformSignZa7,
    CuneiformSignZuOverZuPlusSar,
    CuneiformSignZu5TimesThreeDishTenu,
}

An enum to represent all characters in the EarlyDynasticCuneiform block.

Variants
CuneiformSignAbTimesNunTenu

\u{12480}: 'ğ’’€'

CuneiformSignAbTimesShu2

\u{12481}: 'ğ’’'

CuneiformSignAdTimesEsh2

\u{12482}: 'ğ’’‚'

CuneiformSignBadTimesDishTenu

\u{12483}: 'ğ’’ƒ'

CuneiformSignBahar2TimesAb2

\u{12484}: 'ğ’’„'

CuneiformSignBahar2TimesNi

\u{12485}: 'ğ’’…'

CuneiformSignBahar2TimesZa

\u{12486}: 'ğ’’†'

CuneiformSignBuOverBuTimesNa2

\u{12487}: 'ğ’’‡'

CuneiformSignDaTimesTak4

\u{12488}: 'ğ’’ˆ'

CuneiformSignDagTimesKur

\u{12489}: 'ğ’’‰'

CuneiformSignDimTimesIgi

\u{1248a}: 'ğ’’Š'

CuneiformSignDimTimesUUU

\u{1248b}: 'ğ’’‹'

CuneiformSignDim2TimesUd

\u{1248c}: 'ğ’’Œ'

CuneiformSignDugTimesAnshe

\u{1248d}: 'ğ’’'

CuneiformSignDugTimesAsh

\u{1248e}: 'ğ’’'

CuneiformSignDugTimesAshAtLeft

\u{1248f}: 'ğ’’'

CuneiformSignDugTimesDin

\u{12490}: 'ğ’’'

CuneiformSignDugTimesDun

\u{12491}: 'ğ’’‘'

CuneiformSignDugTimesErin2

\u{12492}: 'ğ’’’'

CuneiformSignDugTimesGa

\u{12493}: 'ğ’’“'

CuneiformSignDugTimesGi

\u{12494}: 'ğ’’”'

CuneiformSignDugTimesGir2Gunu

\u{12495}: 'ğ’’•'

CuneiformSignDugTimesGish

\u{12496}: 'ğ’’–'

CuneiformSignDugTimesHa

\u{12497}: 'ğ’’—'

CuneiformSignDugTimesHi

\u{12498}: 'ğ’’˜'

CuneiformSignDugTimesIgiGunu

\u{12499}: 'ğ’’™'

CuneiformSignDugTimesKaskal

\u{1249a}: 'ğ’’š'

CuneiformSignDugTimesKur

\u{1249b}: 'ğ’’›'

CuneiformSignDugTimesKushu2

\u{1249c}: 'ğ’’œ'

CuneiformSignDugTimesKushu2PlusKaskal

\u{1249d}: 'ğ’’'

CuneiformSignDugTimesLakDash020

\u{1249e}: 'ğ’’'

CuneiformSignDugTimesLam

\u{1249f}: 'ğ’’Ÿ'

CuneiformSignDugTimesLamTimesKur

\u{124a0}: 'ğ’’ '

CuneiformSignDugTimesLuhPlusGish

\u{124a1}: 'ğ’’¡'

CuneiformSignDugTimesMash

\u{124a2}: 'ğ’’¢'

CuneiformSignDugTimesMes

\u{124a3}: 'ğ’’£'

CuneiformSignDugTimesMi

\u{124a4}: 'ğ’’¤'

CuneiformSignDugTimesNi

\u{124a5}: 'ğ’’¥'

CuneiformSignDugTimesPi

\u{124a6}: 'ğ’’¦'

CuneiformSignDugTimesShe

\u{124a7}: 'ğ’’'

CuneiformSignDugTimesSiGunu

\u{124a8}: 'ğ’’¨'

CuneiformSignE2TimesKur

\u{124a9}: 'ğ’’©'

CuneiformSignE2TimesPap

\u{124aa}: 'ğ’’ª'

CuneiformSignErin2X

\u{124ab}: 'ğ’’«'

CuneiformSignEsh2CrossingEsh2

\u{124ac}: 'ğ’’¬'

CuneiformSignEzenSheshigTimesAsh

\u{124ad}: 'ğ’’­'

CuneiformSignEzenSheshigTimesHi

\u{124ae}: 'ğ’’®'

CuneiformSignEzenSheshigTimesIgiGunu

\u{124af}: 'ğ’’¯'

CuneiformSignEzenSheshigTimesLa

\u{124b0}: 'ğ’’°'

CuneiformSignEzenSheshigTimesLal

\u{124b1}: 'ğ’’±'

CuneiformSignEzenSheshigTimesMe

\u{124b2}: 'ğ’’²'

CuneiformSignEzenSheshigTimesMes

\u{124b3}: 'ğ’’³'

CuneiformSignEzenSheshigTimesSu

\u{124b4}: 'ğ’’´'

CuneiformSignEzenTimesSu

\u{124b5}: 'ğ’’µ'

CuneiformSignGa2TimesBahar2

\u{124b6}: 'ğ’’¶'

CuneiformSignGa2TimesDimGunu

\u{124b7}: 'ğ’’·'

CuneiformSignGa2TimesDugTimesIgiGunu

\u{124b8}: 'ğ’’¸'

CuneiformSignGa2TimesDugTimesKaskal

\u{124b9}: 'ğ’’¹'

CuneiformSignGa2TimesEren

\u{124ba}: 'ğ’’º'

CuneiformSignGa2TimesGa

\u{124bb}: 'ğ’’»'

CuneiformSignGa2TimesGarPlusDi

\u{124bc}: 'ğ’’¼'

CuneiformSignGa2TimesGarPlusNe

\u{124bd}: 'ğ’’½'

CuneiformSignGa2TimesHaPlusA

\u{124be}: 'ğ’’¾'

CuneiformSignGa2TimesKushu2PlusKaskal

\u{124bf}: 'ğ’’¿'

CuneiformSignGa2TimesLam

\u{124c0}: 'ğ’“€'

CuneiformSignGa2TimesLamTimesKur

\u{124c1}: 'ğ’“'

CuneiformSignGa2TimesLuh

\u{124c2}: 'ğ’“‚'

CuneiformSignGa2TimesMush

\u{124c3}: 'ğ’“ƒ'

CuneiformSignGa2TimesNe

\u{124c4}: 'ğ’“„'

CuneiformSignGa2TimesNePlusE2

\u{124c5}: 'ğ’“…'

CuneiformSignGa2TimesNePlusGi

\u{124c6}: 'ğ’“†'

CuneiformSignGa2TimesShim

\u{124c7}: 'ğ’“‡'

CuneiformSignGa2TimesZiz2

\u{124c8}: 'ğ’“ˆ'

CuneiformSignGabaRotatedNinetyDegrees

\u{124c9}: 'ğ’“‰'

CuneiformSignGeshtinTimesU

\u{124ca}: 'ğ’“Š'

CuneiformSignGishTimesGishCrossingGish

\u{124cb}: 'ğ’“‹'

CuneiformSignGu2TimesIgiGunu

\u{124cc}: 'ğ’“Œ'

CuneiformSignGudPlusGishTimesTak4

\u{124cd}: 'ğ’“'

CuneiformSignHaTenuGunu

\u{124ce}: 'ğ’“'

CuneiformSignHiTimesAshOverHiTimesAsh

\u{124cf}: 'ğ’“'

CuneiformSignKaTimesBu

\u{124d0}: 'ğ’“'

CuneiformSignKaTimesKa

\u{124d1}: 'ğ’“‘'

CuneiformSignKaTimesUUU

\u{124d2}: 'ğ’“’'

CuneiformSignKaTimesUr

\u{124d3}: 'ğ’““'

CuneiformSignLagabTimesZuOverZu

\u{124d4}: 'ğ’“”'

CuneiformSignLakDash003

\u{124d5}: 'ğ’“•'

CuneiformSignLakDash021

\u{124d6}: 'ğ’“–'

CuneiformSignLakDash025

\u{124d7}: 'ğ’“—'

CuneiformSignLakDash030

\u{124d8}: 'ğ’“˜'

CuneiformSignLakDash050

\u{124d9}: 'ğ’“™'

CuneiformSignLakDash051

\u{124da}: 'ğ’“š'

CuneiformSignLakDash062

\u{124db}: 'ğ’“›'

CuneiformSignLakDash079OverLakDash079Gunu

\u{124dc}: 'ğ’“œ'

CuneiformSignLakDash080

\u{124dd}: 'ğ’“'

CuneiformSignLakDash081OverLakDash081

\u{124de}: 'ğ’“'

CuneiformSignLakDash092

\u{124df}: 'ğ’“Ÿ'

CuneiformSignLakDash130

\u{124e0}: 'ğ’“ '

CuneiformSignLakDash142

\u{124e1}: 'ğ’“¡'

CuneiformSignLakDash210

\u{124e2}: 'ğ’“¢'

CuneiformSignLakDash219

\u{124e3}: 'ğ’“£'

CuneiformSignLakDash220

\u{124e4}: 'ğ’“¤'

CuneiformSignLakDash225

\u{124e5}: 'ğ’“¥'

CuneiformSignLakDash228

\u{124e6}: 'ğ’“¦'

CuneiformSignLakDash238

\u{124e7}: 'ğ’“'

CuneiformSignLakDash265

\u{124e8}: 'ğ’“¨'

CuneiformSignLakDash266

\u{124e9}: 'ğ’“©'

CuneiformSignLakDash343

\u{124ea}: 'ğ’“ª'

CuneiformSignLakDash347

\u{124eb}: 'ğ’“«'

CuneiformSignLakDash348

\u{124ec}: 'ğ’“¬'

CuneiformSignLakDash383

\u{124ed}: 'ğ’“­'

CuneiformSignLakDash384

\u{124ee}: 'ğ’“®'

CuneiformSignLakDash390

\u{124ef}: 'ğ’“¯'

CuneiformSignLakDash441

\u{124f0}: 'ğ’“°'

CuneiformSignLakDash449

\u{124f1}: 'ğ’“±'

CuneiformSignLakDash449TimesGu

\u{124f2}: 'ğ’“²'

CuneiformSignLakDash449TimesIgi

\u{124f3}: 'ğ’“³'

CuneiformSignLakDash449TimesPapPlusLu3

\u{124f4}: 'ğ’“´'

CuneiformSignLakDash449TimesPapPlusPapPlusLu3

\u{124f5}: 'ğ’“µ'

CuneiformSignLakDash449TimesU2PlusBa

\u{124f6}: 'ğ’“¶'

CuneiformSignLakDash450

\u{124f7}: 'ğ’“·'

CuneiformSignLakDash457

\u{124f8}: 'ğ’“¸'

CuneiformSignLakDash470

\u{124f9}: 'ğ’“¹'

CuneiformSignLakDash483

\u{124fa}: 'ğ’“º'

CuneiformSignLakDash490

\u{124fb}: 'ğ’“»'

CuneiformSignLakDash492

\u{124fc}: 'ğ’“¼'

CuneiformSignLakDash493

\u{124fd}: 'ğ’“½'

CuneiformSignLakDash495

\u{124fe}: 'ğ’“¾'

CuneiformSignLakDash550

\u{124ff}: 'ğ’“¿'

CuneiformSignLakDash608

\u{12500}: 'ğ’”€'

CuneiformSignLakDash617

\u{12501}: 'ğ’”'

CuneiformSignLakDash617TimesAsh

\u{12502}: 'ğ’”‚'

CuneiformSignLakDash617TimesBad

\u{12503}: 'ğ’”ƒ'

CuneiformSignLakDash617TimesDun3GunuGunu

\u{12504}: 'ğ’”„'

CuneiformSignLakDash617TimesKu3

\u{12505}: 'ğ’”…'

CuneiformSignLakDash617TimesLa

\u{12506}: 'ğ’”†'

CuneiformSignLakDash617TimesTar

\u{12507}: 'ğ’”‡'

CuneiformSignLakDash617TimesTe

\u{12508}: 'ğ’”ˆ'

CuneiformSignLakDash617TimesU2

\u{12509}: 'ğ’”‰'

CuneiformSignLakDash617TimesUd

\u{1250a}: 'ğ’”Š'

CuneiformSignLakDash617TimesUruda

\u{1250b}: 'ğ’”‹'

CuneiformSignLakDash636

\u{1250c}: 'ğ’”Œ'

CuneiformSignLakDash648

\u{1250d}: 'ğ’”'

CuneiformSignLakDash648TimesDub

\u{1250e}: 'ğ’”'

CuneiformSignLakDash648TimesGa

\u{1250f}: 'ğ’”'

CuneiformSignLakDash648TimesIgi

\u{12510}: 'ğ’”'

CuneiformSignLakDash648TimesIgiGunu

\u{12511}: 'ğ’”‘'

CuneiformSignLakDash648TimesNi

\u{12512}: 'ğ’”’'

CuneiformSignLakDash648TimesPapPlusPapPlusLu3

\u{12513}: 'ğ’”“'

CuneiformSignLakDash648TimesSheshPlusKi

\u{12514}: 'ğ’””'

CuneiformSignLakDash648TimesUd

\u{12515}: 'ğ’”•'

CuneiformSignLakDash648TimesUruda

\u{12516}: 'ğ’”–'

CuneiformSignLakDash724

\u{12517}: 'ğ’”—'

CuneiformSignLakDash749

\u{12518}: 'ğ’”˜'

CuneiformSignLu2GunuTimesAsh

\u{12519}: 'ğ’”™'

CuneiformSignLu2TimesDish

\u{1251a}: 'ğ’”š'

CuneiformSignLu2TimesHal

\u{1251b}: 'ğ’”›'

CuneiformSignLu2TimesPap

\u{1251c}: 'ğ’”œ'

CuneiformSignLu2TimesPapPlusPapPlusLu3

\u{1251d}: 'ğ’”'

CuneiformSignLu2TimesTak4

\u{1251e}: 'ğ’”'

CuneiformSignMiPlusZa7

\u{1251f}: 'ğ’”Ÿ'

CuneiformSignMushOverMushTimesGa

\u{12520}: 'ğ’” '

CuneiformSignMushOverMushTimesKak

\u{12521}: 'ğ’”¡'

CuneiformSignNinda2TimesDimGunu

\u{12522}: 'ğ’”¢'

CuneiformSignNinda2TimesGish

\u{12523}: 'ğ’”£'

CuneiformSignNinda2TimesGul

\u{12524}: 'ğ’”¤'

CuneiformSignNinda2TimesHi

\u{12525}: 'ğ’”¥'

CuneiformSignNinda2TimesKesh2

\u{12526}: 'ğ’”¦'

CuneiformSignNinda2TimesLakDash050

\u{12527}: 'ğ’”'

CuneiformSignNinda2TimesMash

\u{12528}: 'ğ’”¨'

CuneiformSignNinda2TimesPapPlusPap

\u{12529}: 'ğ’”©'

CuneiformSignNinda2TimesU

\u{1252a}: 'ğ’”ª'

CuneiformSignNinda2TimesUPlusU

\u{1252b}: 'ğ’”«'

CuneiformSignNinda2TimesUruda

\u{1252c}: 'ğ’”¬'

CuneiformSignSagGunuTimesHa

\u{1252d}: 'ğ’”­'

CuneiformSignSagTimesEn

\u{1252e}: 'ğ’”®'

CuneiformSignSagTimesSheAtLeft

\u{1252f}: 'ğ’”¯'

CuneiformSignSagTimesTak4

\u{12530}: 'ğ’”°'

CuneiformSignSha6Tenu

\u{12531}: 'ğ’”±'

CuneiformSignSheOverShe

\u{12532}: 'ğ’”²'

CuneiformSignShePlusHub2

\u{12533}: 'ğ’”³'

CuneiformSignShePlusNam2

\u{12534}: 'ğ’”´'

CuneiformSignShePlusSar

\u{12535}: 'ğ’”µ'

CuneiformSignShu2PlusDugTimesNi

\u{12536}: 'ğ’”¶'

CuneiformSignShu2PlusE2TimesAn

\u{12537}: 'ğ’”·'

CuneiformSignSiTimesTak4

\u{12538}: 'ğ’”¸'

CuneiformSignTak4PlusSag

\u{12539}: 'ğ’”¹'

CuneiformSignTumTimesGan2Tenu

\u{1253a}: 'ğ’”º'

CuneiformSignTumTimesThreeDish

\u{1253b}: 'ğ’”»'

CuneiformSignUr2Inverted

\u{1253c}: 'ğ’”¼'

CuneiformSignUr2TimesUd

\u{1253d}: 'ğ’”½'

CuneiformSignUruTimesDara3

\u{1253e}: 'ğ’”¾'

CuneiformSignUruTimesLakDash668

\u{1253f}: 'ğ’”¿'

CuneiformSignUruTimesLu3

\u{12540}: 'ğ’•€'

CuneiformSignZa7

\u{12541}: 'ğ’•'

CuneiformSignZuOverZuPlusSar

\u{12542}: 'ğ’•‚'

CuneiformSignZu5TimesThreeDishTenu

\u{12543}: 'ğ’•ƒ'

Methods

impl EarlyDynasticCuneiform

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case