Enum unicode_types::enclosed_alphanumeric_supplement::EnclosedAlphanumericSupplement

pub enum EnclosedAlphanumericSupplement {
    DigitZeroFullStop,
    DigitZeroComma,
    DigitOneComma,
    DigitTwoComma,
    DigitThreeComma,
    DigitFourComma,
    DigitFiveComma,
    DigitSixComma,
    DigitSevenComma,
    DigitEightComma,
    DigitNineComma,
    DingbatCircledSansDashSerifDigitZero,
    DingbatNegativeCircledSansDashSerifDigitZero,
    ParenthesizedLatinCapitalLetterA,
    ParenthesizedLatinCapitalLetterB,
    ParenthesizedLatinCapitalLetterC,
    ParenthesizedLatinCapitalLetterD,
    ParenthesizedLatinCapitalLetterE,
    ParenthesizedLatinCapitalLetterF,
    ParenthesizedLatinCapitalLetterG,
    ParenthesizedLatinCapitalLetterH,
    ParenthesizedLatinCapitalLetterI,
    ParenthesizedLatinCapitalLetterJ,
    ParenthesizedLatinCapitalLetterK,
    ParenthesizedLatinCapitalLetterL,
    ParenthesizedLatinCapitalLetterM,
    ParenthesizedLatinCapitalLetterN,
    ParenthesizedLatinCapitalLetterO,
    ParenthesizedLatinCapitalLetterP,
    ParenthesizedLatinCapitalLetterQ,
    ParenthesizedLatinCapitalLetterR,
    ParenthesizedLatinCapitalLetterS,
    ParenthesizedLatinCapitalLetterT,
    ParenthesizedLatinCapitalLetterU,
    ParenthesizedLatinCapitalLetterV,
    ParenthesizedLatinCapitalLetterW,
    ParenthesizedLatinCapitalLetterX,
    ParenthesizedLatinCapitalLetterY,
    ParenthesizedLatinCapitalLetterZ,
    TortoiseShellBracketedLatinCapitalLetterS,
    CircledItalicLatinCapitalLetterC,
    CircledItalicLatinCapitalLetterR,
    CircledCd,
    CircledWz,
    CopyleftSymbol,
    SquaredLatinCapitalLetterA,
    SquaredLatinCapitalLetterB,
    SquaredLatinCapitalLetterC,
    SquaredLatinCapitalLetterD,
    SquaredLatinCapitalLetterE,
    SquaredLatinCapitalLetterF,
    SquaredLatinCapitalLetterG,
    SquaredLatinCapitalLetterH,
    SquaredLatinCapitalLetterI,
    SquaredLatinCapitalLetterJ,
    SquaredLatinCapitalLetterK,
    SquaredLatinCapitalLetterL,
    SquaredLatinCapitalLetterM,
    SquaredLatinCapitalLetterN,
    SquaredLatinCapitalLetterO,
    SquaredLatinCapitalLetterP,
    SquaredLatinCapitalLetterQ,
    SquaredLatinCapitalLetterR,
    SquaredLatinCapitalLetterS,
    SquaredLatinCapitalLetterT,
    SquaredLatinCapitalLetterU,
    SquaredLatinCapitalLetterV,
    SquaredLatinCapitalLetterW,
    SquaredLatinCapitalLetterX,
    SquaredLatinCapitalLetterY,
    SquaredLatinCapitalLetterZ,
    SquaredHv,
    SquaredMv,
    SquaredSd,
    SquaredSs,
    SquaredPpv,
    SquaredWc,
    NegativeCircledLatinCapitalLetterA,
    NegativeCircledLatinCapitalLetterB,
    NegativeCircledLatinCapitalLetterC,
    NegativeCircledLatinCapitalLetterD,
    NegativeCircledLatinCapitalLetterE,
    NegativeCircledLatinCapitalLetterF,
    NegativeCircledLatinCapitalLetterG,
    NegativeCircledLatinCapitalLetterH,
    NegativeCircledLatinCapitalLetterI,
    NegativeCircledLatinCapitalLetterJ,
    NegativeCircledLatinCapitalLetterK,
    NegativeCircledLatinCapitalLetterL,
    NegativeCircledLatinCapitalLetterM,
    NegativeCircledLatinCapitalLetterN,
    NegativeCircledLatinCapitalLetterO,
    NegativeCircledLatinCapitalLetterP,
    NegativeCircledLatinCapitalLetterQ,
    NegativeCircledLatinCapitalLetterR,
    NegativeCircledLatinCapitalLetterS,
    NegativeCircledLatinCapitalLetterT,
    NegativeCircledLatinCapitalLetterU,
    NegativeCircledLatinCapitalLetterV,
    NegativeCircledLatinCapitalLetterW,
    NegativeCircledLatinCapitalLetterX,
    NegativeCircledLatinCapitalLetterY,
    NegativeCircledLatinCapitalLetterZ,
    RaisedMcSign,
    RaisedMdSign,
    RaisedMrSign,
    NegativeSquaredLatinCapitalLetterA,
    NegativeSquaredLatinCapitalLetterB,
    NegativeSquaredLatinCapitalLetterC,
    NegativeSquaredLatinCapitalLetterD,
    NegativeSquaredLatinCapitalLetterE,
    NegativeSquaredLatinCapitalLetterF,
    NegativeSquaredLatinCapitalLetterG,
    NegativeSquaredLatinCapitalLetterH,
    NegativeSquaredLatinCapitalLetterI,
    NegativeSquaredLatinCapitalLetterJ,
    NegativeSquaredLatinCapitalLetterK,
    NegativeSquaredLatinCapitalLetterL,
    NegativeSquaredLatinCapitalLetterM,
    NegativeSquaredLatinCapitalLetterN,
    NegativeSquaredLatinCapitalLetterO,
    NegativeSquaredLatinCapitalLetterP,
    NegativeSquaredLatinCapitalLetterQ,
    NegativeSquaredLatinCapitalLetterR,
    NegativeSquaredLatinCapitalLetterS,
    NegativeSquaredLatinCapitalLetterT,
    NegativeSquaredLatinCapitalLetterU,
    NegativeSquaredLatinCapitalLetterV,
    NegativeSquaredLatinCapitalLetterW,
    NegativeSquaredLatinCapitalLetterX,
    NegativeSquaredLatinCapitalLetterY,
    NegativeSquaredLatinCapitalLetterZ,
    CrossedNegativeSquaredLatinCapitalLetterP,
    NegativeSquaredIc,
    NegativeSquaredPa,
    NegativeSquaredSa,
    NegativeSquaredAb,
    NegativeSquaredWc,
    SquareDj,
    SquaredCl,
    SquaredCool,
    SquaredFree,
    SquaredId,
    SquaredNew,
    SquaredNg,
    SquaredOk,
    SquaredSos,
    SquaredUpWithExclamationMark,
    SquaredVs,
    SquaredThreeD,
    SquaredSecondScreen,
    SquaredTwoK,
    SquaredFourK,
    SquaredEightK,
    SquaredFivePointOne,
    SquaredSevenPointOne,
    SquaredTwentyDashTwoPointTwo,
    SquaredSixtyP,
    SquaredOneHundredTwentyP,
    SquaredLatinSmallLetterD,
    SquaredHc,
    SquaredHdr,
    SquaredHiDashRes,
    SquaredLossless,
    SquaredShv,
    SquaredUhd,
    SquaredVod,
    RegionalIndicatorSymbolLetterA,
    RegionalIndicatorSymbolLetterB,
    RegionalIndicatorSymbolLetterC,
    RegionalIndicatorSymbolLetterD,
    RegionalIndicatorSymbolLetterE,
    RegionalIndicatorSymbolLetterF,
    RegionalIndicatorSymbolLetterG,
    RegionalIndicatorSymbolLetterH,
    RegionalIndicatorSymbolLetterI,
    RegionalIndicatorSymbolLetterJ,
    RegionalIndicatorSymbolLetterK,
    RegionalIndicatorSymbolLetterL,
    RegionalIndicatorSymbolLetterM,
    RegionalIndicatorSymbolLetterN,
    RegionalIndicatorSymbolLetterO,
    RegionalIndicatorSymbolLetterP,
    RegionalIndicatorSymbolLetterQ,
    RegionalIndicatorSymbolLetterR,
    RegionalIndicatorSymbolLetterS,
    RegionalIndicatorSymbolLetterT,
    RegionalIndicatorSymbolLetterU,
    RegionalIndicatorSymbolLetterV,
    RegionalIndicatorSymbolLetterW,
    RegionalIndicatorSymbolLetterX,
    RegionalIndicatorSymbolLetterY,
}

An enum to represent all characters in the EnclosedAlphanumericSupplement block.

Variants
DigitZeroFullStop

\u{1f100}: 'ğŸ„€'

DigitZeroComma

\u{1f101}: 'ğŸ„'

DigitOneComma

\u{1f102}: 'ğŸ„‚'

DigitTwoComma

\u{1f103}: 'ğŸ„ƒ'

DigitThreeComma

\u{1f104}: 'ğŸ„„'

DigitFourComma

\u{1f105}: 'ğŸ„…'

DigitFiveComma

\u{1f106}: 'ğŸ„†'

DigitSixComma

\u{1f107}: 'ğŸ„‡'

DigitSevenComma

\u{1f108}: 'ğŸ„ˆ'

DigitEightComma

\u{1f109}: 'ğŸ„‰'

DigitNineComma

\u{1f10a}: 'ğŸ„Š'

DingbatCircledSansDashSerifDigitZero

\u{1f10b}: 'ğŸ„‹'

DingbatNegativeCircledSansDashSerifDigitZero

\u{1f10c}: 'ğŸ„Œ'

ParenthesizedLatinCapitalLetterA

\u{1f110}: 'ğŸ„'

ParenthesizedLatinCapitalLetterB

\u{1f111}: 'ğŸ„‘'

ParenthesizedLatinCapitalLetterC

\u{1f112}: 'ğŸ„’'

ParenthesizedLatinCapitalLetterD

\u{1f113}: 'ğŸ„“'

ParenthesizedLatinCapitalLetterE

\u{1f114}: 'ğŸ„”'

ParenthesizedLatinCapitalLetterF

\u{1f115}: 'ğŸ„•'

ParenthesizedLatinCapitalLetterG

\u{1f116}: 'ğŸ„–'

ParenthesizedLatinCapitalLetterH

\u{1f117}: 'ğŸ„—'

ParenthesizedLatinCapitalLetterI

\u{1f118}: 'ğŸ„˜'

ParenthesizedLatinCapitalLetterJ

\u{1f119}: 'ğŸ„™'

ParenthesizedLatinCapitalLetterK

\u{1f11a}: 'ğŸ„š'

ParenthesizedLatinCapitalLetterL

\u{1f11b}: 'ğŸ„›'

ParenthesizedLatinCapitalLetterM

\u{1f11c}: 'ğŸ„œ'

ParenthesizedLatinCapitalLetterN

\u{1f11d}: 'ğŸ„'

ParenthesizedLatinCapitalLetterO

\u{1f11e}: 'ğŸ„'

ParenthesizedLatinCapitalLetterP

\u{1f11f}: 'ğŸ„Ÿ'

ParenthesizedLatinCapitalLetterQ

\u{1f120}: 'ğŸ„ '

ParenthesizedLatinCapitalLetterR

\u{1f121}: 'ğŸ„¡'

ParenthesizedLatinCapitalLetterS

\u{1f122}: 'ğŸ„¢'

ParenthesizedLatinCapitalLetterT

\u{1f123}: 'ğŸ„£'

ParenthesizedLatinCapitalLetterU

\u{1f124}: 'ğŸ„¤'

ParenthesizedLatinCapitalLetterV

\u{1f125}: 'ğŸ„¥'

ParenthesizedLatinCapitalLetterW

\u{1f126}: 'ğŸ„¦'

ParenthesizedLatinCapitalLetterX

\u{1f127}: 'ğŸ„'

ParenthesizedLatinCapitalLetterY

\u{1f128}: 'ğŸ„¨'

ParenthesizedLatinCapitalLetterZ

\u{1f129}: 'ğŸ„©'

TortoiseShellBracketedLatinCapitalLetterS

\u{1f12a}: 'ğŸ„ª'

CircledItalicLatinCapitalLetterC

\u{1f12b}: 'ğŸ„«'

CircledItalicLatinCapitalLetterR

\u{1f12c}: 'ğŸ„¬'

CircledCd

\u{1f12d}: 'ğŸ„­'

CircledWz

\u{1f12e}: 'ğŸ„®'

CopyleftSymbol

\u{1f12f}: 'ğŸ„¯'

SquaredLatinCapitalLetterA

\u{1f130}: 'ğŸ„°'

SquaredLatinCapitalLetterB

\u{1f131}: 'ğŸ„±'

SquaredLatinCapitalLetterC

\u{1f132}: 'ğŸ„²'

SquaredLatinCapitalLetterD

\u{1f133}: 'ğŸ„³'

SquaredLatinCapitalLetterE

\u{1f134}: 'ğŸ„´'

SquaredLatinCapitalLetterF

\u{1f135}: 'ğŸ„µ'

SquaredLatinCapitalLetterG

\u{1f136}: 'ğŸ„¶'

SquaredLatinCapitalLetterH

\u{1f137}: 'ğŸ„·'

SquaredLatinCapitalLetterI

\u{1f138}: 'ğŸ„¸'

SquaredLatinCapitalLetterJ

\u{1f139}: 'ğŸ„¹'

SquaredLatinCapitalLetterK

\u{1f13a}: 'ğŸ„º'

SquaredLatinCapitalLetterL

\u{1f13b}: 'ğŸ„»'

SquaredLatinCapitalLetterM

\u{1f13c}: 'ğŸ„¼'

SquaredLatinCapitalLetterN

\u{1f13d}: 'ğŸ„½'

SquaredLatinCapitalLetterO

\u{1f13e}: 'ğŸ„¾'

SquaredLatinCapitalLetterP

\u{1f13f}: 'ğŸ„¿'

SquaredLatinCapitalLetterQ

\u{1f140}: 'ğŸ…€'

SquaredLatinCapitalLetterR

\u{1f141}: 'ğŸ…'

SquaredLatinCapitalLetterS

\u{1f142}: 'ğŸ…‚'

SquaredLatinCapitalLetterT

\u{1f143}: 'ğŸ…ƒ'

SquaredLatinCapitalLetterU

\u{1f144}: 'ğŸ…„'

SquaredLatinCapitalLetterV

\u{1f145}: 'ğŸ……'

SquaredLatinCapitalLetterW

\u{1f146}: 'ğŸ…†'

SquaredLatinCapitalLetterX

\u{1f147}: 'ğŸ…‡'

SquaredLatinCapitalLetterY

\u{1f148}: 'ğŸ…ˆ'

SquaredLatinCapitalLetterZ

\u{1f149}: 'ğŸ…‰'

SquaredHv

\u{1f14a}: 'ğŸ…Š'

SquaredMv

\u{1f14b}: 'ğŸ…‹'

SquaredSd

\u{1f14c}: 'ğŸ…Œ'

SquaredSs

\u{1f14d}: 'ğŸ…'

SquaredPpv

\u{1f14e}: 'ğŸ…'

SquaredWc

\u{1f14f}: 'ğŸ…'

NegativeCircledLatinCapitalLetterA

\u{1f150}: 'ğŸ…'

NegativeCircledLatinCapitalLetterB

\u{1f151}: 'ğŸ…‘'

NegativeCircledLatinCapitalLetterC

\u{1f152}: 'ğŸ…’'

NegativeCircledLatinCapitalLetterD

\u{1f153}: 'ğŸ…“'

NegativeCircledLatinCapitalLetterE

\u{1f154}: 'ğŸ…”'

NegativeCircledLatinCapitalLetterF

\u{1f155}: 'ğŸ…•'

NegativeCircledLatinCapitalLetterG

\u{1f156}: 'ğŸ…–'

NegativeCircledLatinCapitalLetterH

\u{1f157}: 'ğŸ…—'

NegativeCircledLatinCapitalLetterI

\u{1f158}: 'ğŸ…˜'

NegativeCircledLatinCapitalLetterJ

\u{1f159}: 'ğŸ…™'

NegativeCircledLatinCapitalLetterK

\u{1f15a}: 'ğŸ…š'

NegativeCircledLatinCapitalLetterL

\u{1f15b}: 'ğŸ…›'

NegativeCircledLatinCapitalLetterM

\u{1f15c}: 'ğŸ…œ'

NegativeCircledLatinCapitalLetterN

\u{1f15d}: 'ğŸ…'

NegativeCircledLatinCapitalLetterO

\u{1f15e}: 'ğŸ…'

NegativeCircledLatinCapitalLetterP

\u{1f15f}: 'ğŸ…Ÿ'

NegativeCircledLatinCapitalLetterQ

\u{1f160}: 'ğŸ… '

NegativeCircledLatinCapitalLetterR

\u{1f161}: 'ğŸ…¡'

NegativeCircledLatinCapitalLetterS

\u{1f162}: 'ğŸ…¢'

NegativeCircledLatinCapitalLetterT

\u{1f163}: 'ğŸ…£'

NegativeCircledLatinCapitalLetterU

\u{1f164}: 'ğŸ…¤'

NegativeCircledLatinCapitalLetterV

\u{1f165}: 'ğŸ…¥'

NegativeCircledLatinCapitalLetterW

\u{1f166}: 'ğŸ…¦'

NegativeCircledLatinCapitalLetterX

\u{1f167}: 'ğŸ…'

NegativeCircledLatinCapitalLetterY

\u{1f168}: 'ğŸ…¨'

NegativeCircledLatinCapitalLetterZ

\u{1f169}: 'ğŸ…©'

RaisedMcSign

\u{1f16a}: 'ğŸ…ª'

RaisedMdSign

\u{1f16b}: 'ğŸ…«'

RaisedMrSign

\u{1f16c}: 'ğŸ…¬'

NegativeSquaredLatinCapitalLetterA

\u{1f170}: 'ğŸ…°'

NegativeSquaredLatinCapitalLetterB

\u{1f171}: 'ğŸ…±'

NegativeSquaredLatinCapitalLetterC

\u{1f172}: 'ğŸ…²'

NegativeSquaredLatinCapitalLetterD

\u{1f173}: 'ğŸ…³'

NegativeSquaredLatinCapitalLetterE

\u{1f174}: 'ğŸ…´'

NegativeSquaredLatinCapitalLetterF

\u{1f175}: 'ğŸ…µ'

NegativeSquaredLatinCapitalLetterG

\u{1f176}: 'ğŸ…¶'

NegativeSquaredLatinCapitalLetterH

\u{1f177}: 'ğŸ…·'

NegativeSquaredLatinCapitalLetterI

\u{1f178}: 'ğŸ…¸'

NegativeSquaredLatinCapitalLetterJ

\u{1f179}: 'ğŸ…¹'

NegativeSquaredLatinCapitalLetterK

\u{1f17a}: 'ğŸ…º'

NegativeSquaredLatinCapitalLetterL

\u{1f17b}: 'ğŸ…»'

NegativeSquaredLatinCapitalLetterM

\u{1f17c}: 'ğŸ…¼'

NegativeSquaredLatinCapitalLetterN

\u{1f17d}: 'ğŸ…½'

NegativeSquaredLatinCapitalLetterO

\u{1f17e}: 'ğŸ…¾'

NegativeSquaredLatinCapitalLetterP

\u{1f17f}: 'ğŸ…¿'

NegativeSquaredLatinCapitalLetterQ

\u{1f180}: 'ğŸ†€'

NegativeSquaredLatinCapitalLetterR

\u{1f181}: 'ğŸ†'

NegativeSquaredLatinCapitalLetterS

\u{1f182}: 'ğŸ†‚'

NegativeSquaredLatinCapitalLetterT

\u{1f183}: 'ğŸ†ƒ'

NegativeSquaredLatinCapitalLetterU

\u{1f184}: 'ğŸ†„'

NegativeSquaredLatinCapitalLetterV

\u{1f185}: 'ğŸ†…'

NegativeSquaredLatinCapitalLetterW

\u{1f186}: 'ğŸ††'

NegativeSquaredLatinCapitalLetterX

\u{1f187}: 'ğŸ†‡'

NegativeSquaredLatinCapitalLetterY

\u{1f188}: 'ğŸ†ˆ'

NegativeSquaredLatinCapitalLetterZ

\u{1f189}: 'ğŸ†‰'

CrossedNegativeSquaredLatinCapitalLetterP

\u{1f18a}: 'ğŸ†Š'

NegativeSquaredIc

\u{1f18b}: 'ğŸ†‹'

NegativeSquaredPa

\u{1f18c}: 'ğŸ†Œ'

NegativeSquaredSa

\u{1f18d}: 'ğŸ†'

NegativeSquaredAb

\u{1f18e}: 'ğŸ†'

NegativeSquaredWc

\u{1f18f}: 'ğŸ†'

SquareDj

\u{1f190}: 'ğŸ†'

SquaredCl

\u{1f191}: 'ğŸ†‘'

SquaredCool

\u{1f192}: 'ğŸ†’'

SquaredFree

\u{1f193}: 'ğŸ†“'

SquaredId

\u{1f194}: 'ğŸ†”'

SquaredNew

\u{1f195}: 'ğŸ†•'

SquaredNg

\u{1f196}: 'ğŸ†–'

SquaredOk

\u{1f197}: 'ğŸ†—'

SquaredSos

\u{1f198}: 'ğŸ†˜'

SquaredUpWithExclamationMark

\u{1f199}: 'ğŸ†™'

SquaredVs

\u{1f19a}: 'ğŸ†š'

SquaredThreeD

\u{1f19b}: 'ğŸ†›'

SquaredSecondScreen

\u{1f19c}: 'ğŸ†œ'

SquaredTwoK

\u{1f19d}: 'ğŸ†'

SquaredFourK

\u{1f19e}: 'ğŸ†'

SquaredEightK

\u{1f19f}: 'ğŸ†Ÿ'

SquaredFivePointOne

\u{1f1a0}: 'ğŸ† '

SquaredSevenPointOne

\u{1f1a1}: 'ğŸ†¡'

SquaredTwentyDashTwoPointTwo

\u{1f1a2}: 'ğŸ†¢'

SquaredSixtyP

\u{1f1a3}: 'ğŸ†£'

SquaredOneHundredTwentyP

\u{1f1a4}: 'ğŸ†¤'

SquaredLatinSmallLetterD

\u{1f1a5}: 'ğŸ†¥'

SquaredHc

\u{1f1a6}: 'ğŸ†¦'

SquaredHdr

\u{1f1a7}: 'ğŸ†'

SquaredHiDashRes

\u{1f1a8}: 'ğŸ†¨'

SquaredLossless

\u{1f1a9}: 'ğŸ†©'

SquaredShv

\u{1f1aa}: 'ğŸ†ª'

SquaredUhd

\u{1f1ab}: 'ğŸ†«'

SquaredVod

\u{1f1ac}: 'ğŸ†¬'

RegionalIndicatorSymbolLetterA

\u{1f1e6}: 'ğŸ‡¦'

RegionalIndicatorSymbolLetterB

\u{1f1e7}: 'ğŸ‡'

RegionalIndicatorSymbolLetterC

\u{1f1e8}: 'ğŸ‡¨'

RegionalIndicatorSymbolLetterD

\u{1f1e9}: 'ğŸ‡©'

RegionalIndicatorSymbolLetterE

\u{1f1ea}: 'ğŸ‡ª'

RegionalIndicatorSymbolLetterF

\u{1f1eb}: 'ğŸ‡«'

RegionalIndicatorSymbolLetterG

\u{1f1ec}: 'ğŸ‡¬'

RegionalIndicatorSymbolLetterH

\u{1f1ed}: 'ğŸ‡­'

RegionalIndicatorSymbolLetterI

\u{1f1ee}: 'ğŸ‡®'

RegionalIndicatorSymbolLetterJ

\u{1f1ef}: 'ğŸ‡¯'

RegionalIndicatorSymbolLetterK

\u{1f1f0}: 'ğŸ‡°'

RegionalIndicatorSymbolLetterL

\u{1f1f1}: 'ğŸ‡±'

RegionalIndicatorSymbolLetterM

\u{1f1f2}: 'ğŸ‡²'

RegionalIndicatorSymbolLetterN

\u{1f1f3}: 'ğŸ‡³'

RegionalIndicatorSymbolLetterO

\u{1f1f4}: 'ğŸ‡´'

RegionalIndicatorSymbolLetterP

\u{1f1f5}: 'ğŸ‡µ'

RegionalIndicatorSymbolLetterQ

\u{1f1f6}: 'ğŸ‡¶'

RegionalIndicatorSymbolLetterR

\u{1f1f7}: 'ğŸ‡·'

RegionalIndicatorSymbolLetterS

\u{1f1f8}: 'ğŸ‡¸'

RegionalIndicatorSymbolLetterT

\u{1f1f9}: 'ğŸ‡¹'

RegionalIndicatorSymbolLetterU

\u{1f1fa}: 'ğŸ‡º'

RegionalIndicatorSymbolLetterV

\u{1f1fb}: 'ğŸ‡»'

RegionalIndicatorSymbolLetterW

\u{1f1fc}: 'ğŸ‡¼'

RegionalIndicatorSymbolLetterX

\u{1f1fd}: 'ğŸ‡½'

RegionalIndicatorSymbolLetterY

\u{1f1fe}: 'ğŸ‡¾'

Methods

impl EnclosedAlphanumericSupplement

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case