Enum unicode_types::enclosed_alphanumerics::EnclosedAlphanumerics

pub enum EnclosedAlphanumerics {
    CircledDigitOne,
    CircledDigitTwo,
    CircledDigitThree,
    CircledDigitFour,
    CircledDigitFive,
    CircledDigitSix,
    CircledDigitSeven,
    CircledDigitEight,
    CircledDigitNine,
    CircledNumberTen,
    CircledNumberEleven,
    CircledNumberTwelve,
    CircledNumberThirteen,
    CircledNumberFourteen,
    CircledNumberFifteen,
    CircledNumberSixteen,
    CircledNumberSeventeen,
    CircledNumberEighteen,
    CircledNumberNineteen,
    CircledNumberTwenty,
    ParenthesizedDigitOne,
    ParenthesizedDigitTwo,
    ParenthesizedDigitThree,
    ParenthesizedDigitFour,
    ParenthesizedDigitFive,
    ParenthesizedDigitSix,
    ParenthesizedDigitSeven,
    ParenthesizedDigitEight,
    ParenthesizedDigitNine,
    ParenthesizedNumberTen,
    ParenthesizedNumberEleven,
    ParenthesizedNumberTwelve,
    ParenthesizedNumberThirteen,
    ParenthesizedNumberFourteen,
    ParenthesizedNumberFifteen,
    ParenthesizedNumberSixteen,
    ParenthesizedNumberSeventeen,
    ParenthesizedNumberEighteen,
    ParenthesizedNumberNineteen,
    ParenthesizedNumberTwenty,
    DigitOneFullStop,
    DigitTwoFullStop,
    DigitThreeFullStop,
    DigitFourFullStop,
    DigitFiveFullStop,
    DigitSixFullStop,
    DigitSevenFullStop,
    DigitEightFullStop,
    DigitNineFullStop,
    NumberTenFullStop,
    NumberElevenFullStop,
    NumberTwelveFullStop,
    NumberThirteenFullStop,
    NumberFourteenFullStop,
    NumberFifteenFullStop,
    NumberSixteenFullStop,
    NumberSeventeenFullStop,
    NumberEighteenFullStop,
    NumberNineteenFullStop,
    NumberTwentyFullStop,
    ParenthesizedLatinSmallLetterA,
    ParenthesizedLatinSmallLetterB,
    ParenthesizedLatinSmallLetterC,
    ParenthesizedLatinSmallLetterD,
    ParenthesizedLatinSmallLetterE,
    ParenthesizedLatinSmallLetterF,
    ParenthesizedLatinSmallLetterG,
    ParenthesizedLatinSmallLetterH,
    ParenthesizedLatinSmallLetterI,
    ParenthesizedLatinSmallLetterJ,
    ParenthesizedLatinSmallLetterK,
    ParenthesizedLatinSmallLetterL,
    ParenthesizedLatinSmallLetterM,
    ParenthesizedLatinSmallLetterN,
    ParenthesizedLatinSmallLetterO,
    ParenthesizedLatinSmallLetterP,
    ParenthesizedLatinSmallLetterQ,
    ParenthesizedLatinSmallLetterR,
    ParenthesizedLatinSmallLetterS,
    ParenthesizedLatinSmallLetterT,
    ParenthesizedLatinSmallLetterU,
    ParenthesizedLatinSmallLetterV,
    ParenthesizedLatinSmallLetterW,
    ParenthesizedLatinSmallLetterX,
    ParenthesizedLatinSmallLetterY,
    ParenthesizedLatinSmallLetterZ,
    CircledLatinCapitalLetterA,
    CircledLatinCapitalLetterB,
    CircledLatinCapitalLetterC,
    CircledLatinCapitalLetterD,
    CircledLatinCapitalLetterE,
    CircledLatinCapitalLetterF,
    CircledLatinCapitalLetterG,
    CircledLatinCapitalLetterH,
    CircledLatinCapitalLetterI,
    CircledLatinCapitalLetterJ,
    CircledLatinCapitalLetterK,
    CircledLatinCapitalLetterL,
    CircledLatinCapitalLetterM,
    CircledLatinCapitalLetterN,
    CircledLatinCapitalLetterO,
    CircledLatinCapitalLetterP,
    CircledLatinCapitalLetterQ,
    CircledLatinCapitalLetterR,
    CircledLatinCapitalLetterS,
    CircledLatinCapitalLetterT,
    CircledLatinCapitalLetterU,
    CircledLatinCapitalLetterV,
    CircledLatinCapitalLetterW,
    CircledLatinCapitalLetterX,
    CircledLatinCapitalLetterY,
    CircledLatinCapitalLetterZ,
    CircledLatinSmallLetterA,
    CircledLatinSmallLetterB,
    CircledLatinSmallLetterC,
    CircledLatinSmallLetterD,
    CircledLatinSmallLetterE,
    CircledLatinSmallLetterF,
    CircledLatinSmallLetterG,
    CircledLatinSmallLetterH,
    CircledLatinSmallLetterI,
    CircledLatinSmallLetterJ,
    CircledLatinSmallLetterK,
    CircledLatinSmallLetterL,
    CircledLatinSmallLetterM,
    CircledLatinSmallLetterN,
    CircledLatinSmallLetterO,
    CircledLatinSmallLetterP,
    CircledLatinSmallLetterQ,
    CircledLatinSmallLetterR,
    CircledLatinSmallLetterS,
    CircledLatinSmallLetterT,
    CircledLatinSmallLetterU,
    CircledLatinSmallLetterV,
    CircledLatinSmallLetterW,
    CircledLatinSmallLetterX,
    CircledLatinSmallLetterY,
    CircledLatinSmallLetterZ,
    CircledDigitZero,
    NegativeCircledNumberEleven,
    NegativeCircledNumberTwelve,
    NegativeCircledNumberThirteen,
    NegativeCircledNumberFourteen,
    NegativeCircledNumberFifteen,
    NegativeCircledNumberSixteen,
    NegativeCircledNumberSeventeen,
    NegativeCircledNumberEighteen,
    NegativeCircledNumberNineteen,
    NegativeCircledNumberTwenty,
    DoubleCircledDigitOne,
    DoubleCircledDigitTwo,
    DoubleCircledDigitThree,
    DoubleCircledDigitFour,
    DoubleCircledDigitFive,
    DoubleCircledDigitSix,
    DoubleCircledDigitSeven,
    DoubleCircledDigitEight,
    DoubleCircledDigitNine,
    DoubleCircledNumberTen,
}

An enum to represent all characters in the EnclosedAlphanumerics block.

Variants
CircledDigitOne

\u{2460}: '①'

CircledDigitTwo

\u{2461}: '②'

CircledDigitThree

\u{2462}: '③'

CircledDigitFour

\u{2463}: '④'

CircledDigitFive

\u{2464}: '⑤'

CircledDigitSix

\u{2465}: '⑥'

CircledDigitSeven

\u{2466}: '⑦'

CircledDigitEight

\u{2467}: '⑧'

CircledDigitNine

\u{2468}: '⑨'

CircledNumberTen

\u{2469}: '⑩'

CircledNumberEleven

\u{246a}: '⑪'

CircledNumberTwelve

\u{246b}: '⑫'

CircledNumberThirteen

\u{246c}: '⑬'

CircledNumberFourteen

\u{246d}: '⑭'

CircledNumberFifteen

\u{246e}: '⑮'

CircledNumberSixteen

\u{246f}: '⑯'

CircledNumberSeventeen

\u{2470}: '⑰'

CircledNumberEighteen

\u{2471}: '⑱'

CircledNumberNineteen

\u{2472}: '⑲'

CircledNumberTwenty

\u{2473}: '⑳'

ParenthesizedDigitOne

\u{2474}: '⑴'

ParenthesizedDigitTwo

\u{2475}: '⑵'

ParenthesizedDigitThree

\u{2476}: '⑶'

ParenthesizedDigitFour

\u{2477}: '⑷'

ParenthesizedDigitFive

\u{2478}: '⑸'

ParenthesizedDigitSix

\u{2479}: '⑹'

ParenthesizedDigitSeven

\u{247a}: '⑺'

ParenthesizedDigitEight

\u{247b}: '⑻'

ParenthesizedDigitNine

\u{247c}: '⑼'

ParenthesizedNumberTen

\u{247d}: '⑽'

ParenthesizedNumberEleven

\u{247e}: '⑾'

ParenthesizedNumberTwelve

\u{247f}: '⑿'

ParenthesizedNumberThirteen

\u{2480}: '⒀'

ParenthesizedNumberFourteen

\u{2481}: '⒁'

ParenthesizedNumberFifteen

\u{2482}: '⒂'

ParenthesizedNumberSixteen

\u{2483}: '⒃'

ParenthesizedNumberSeventeen

\u{2484}: '⒄'

ParenthesizedNumberEighteen

\u{2485}: '⒅'

ParenthesizedNumberNineteen

\u{2486}: '⒆'

ParenthesizedNumberTwenty

\u{2487}: '⒇'

DigitOneFullStop

\u{2488}: '⒈'

DigitTwoFullStop

\u{2489}: '⒉'

DigitThreeFullStop

\u{248a}: '⒊'

DigitFourFullStop

\u{248b}: '⒋'

DigitFiveFullStop

\u{248c}: '⒌'

DigitSixFullStop

\u{248d}: '⒍'

DigitSevenFullStop

\u{248e}: '⒎'

DigitEightFullStop

\u{248f}: '⒏'

DigitNineFullStop

\u{2490}: '⒐'

NumberTenFullStop

\u{2491}: '⒑'

NumberElevenFullStop

\u{2492}: '⒒'

NumberTwelveFullStop

\u{2493}: '⒓'

NumberThirteenFullStop

\u{2494}: '⒔'

NumberFourteenFullStop

\u{2495}: '⒕'

NumberFifteenFullStop

\u{2496}: '⒖'

NumberSixteenFullStop

\u{2497}: '⒗'

NumberSeventeenFullStop

\u{2498}: '⒘'

NumberEighteenFullStop

\u{2499}: '⒙'

NumberNineteenFullStop

\u{249a}: '⒚'

NumberTwentyFullStop

\u{249b}: '⒛'

ParenthesizedLatinSmallLetterA

\u{249c}: '⒜'

ParenthesizedLatinSmallLetterB

\u{249d}: '⒝'

ParenthesizedLatinSmallLetterC

\u{249e}: '⒞'

ParenthesizedLatinSmallLetterD

\u{249f}: '⒟'

ParenthesizedLatinSmallLetterE

\u{24a0}: '⒠'

ParenthesizedLatinSmallLetterF

\u{24a1}: '⒡'

ParenthesizedLatinSmallLetterG

\u{24a2}: '⒢'

ParenthesizedLatinSmallLetterH

\u{24a3}: '⒣'

ParenthesizedLatinSmallLetterI

\u{24a4}: '⒤'

ParenthesizedLatinSmallLetterJ

\u{24a5}: '⒥'

ParenthesizedLatinSmallLetterK

\u{24a6}: '⒦'

ParenthesizedLatinSmallLetterL

\u{24a7}: '⒧'

ParenthesizedLatinSmallLetterM

\u{24a8}: '⒨'

ParenthesizedLatinSmallLetterN

\u{24a9}: '⒩'

ParenthesizedLatinSmallLetterO

\u{24aa}: '⒪'

ParenthesizedLatinSmallLetterP

\u{24ab}: '⒫'

ParenthesizedLatinSmallLetterQ

\u{24ac}: '⒬'

ParenthesizedLatinSmallLetterR

\u{24ad}: '⒭'

ParenthesizedLatinSmallLetterS

\u{24ae}: '⒮'

ParenthesizedLatinSmallLetterT

\u{24af}: '⒯'

ParenthesizedLatinSmallLetterU

\u{24b0}: '⒰'

ParenthesizedLatinSmallLetterV

\u{24b1}: '⒱'

ParenthesizedLatinSmallLetterW

\u{24b2}: '⒲'

ParenthesizedLatinSmallLetterX

\u{24b3}: '⒳'

ParenthesizedLatinSmallLetterY

\u{24b4}: '⒴'

ParenthesizedLatinSmallLetterZ

\u{24b5}: '⒵'

CircledLatinCapitalLetterA

\u{24b6}: 'Ⓐ'

CircledLatinCapitalLetterB

\u{24b7}: 'Ⓑ'

CircledLatinCapitalLetterC

\u{24b8}: 'Ⓒ'

CircledLatinCapitalLetterD

\u{24b9}: 'Ⓓ'

CircledLatinCapitalLetterE

\u{24ba}: 'Ⓔ'

CircledLatinCapitalLetterF

\u{24bb}: 'Ⓕ'

CircledLatinCapitalLetterG

\u{24bc}: 'Ⓖ'

CircledLatinCapitalLetterH

\u{24bd}: 'Ⓗ'

CircledLatinCapitalLetterI

\u{24be}: 'Ⓘ'

CircledLatinCapitalLetterJ

\u{24bf}: 'Ⓙ'

CircledLatinCapitalLetterK

\u{24c0}: 'Ⓚ'

CircledLatinCapitalLetterL

\u{24c1}: 'Ⓛ'

CircledLatinCapitalLetterM

\u{24c2}: 'Ⓜ'

CircledLatinCapitalLetterN

\u{24c3}: 'Ⓝ'

CircledLatinCapitalLetterO

\u{24c4}: 'Ⓞ'

CircledLatinCapitalLetterP

\u{24c5}: 'Ⓟ'

CircledLatinCapitalLetterQ

\u{24c6}: 'Ⓠ'

CircledLatinCapitalLetterR

\u{24c7}: 'Ⓡ'

CircledLatinCapitalLetterS

\u{24c8}: 'Ⓢ'

CircledLatinCapitalLetterT

\u{24c9}: 'Ⓣ'

CircledLatinCapitalLetterU

\u{24ca}: 'Ⓤ'

CircledLatinCapitalLetterV

\u{24cb}: 'Ⓥ'

CircledLatinCapitalLetterW

\u{24cc}: 'Ⓦ'

CircledLatinCapitalLetterX

\u{24cd}: 'Ⓧ'

CircledLatinCapitalLetterY

\u{24ce}: 'Ⓨ'

CircledLatinCapitalLetterZ

\u{24cf}: 'Ⓩ'

CircledLatinSmallLetterA

\u{24d0}: 'ⓐ'

CircledLatinSmallLetterB

\u{24d1}: 'ⓑ'

CircledLatinSmallLetterC

\u{24d2}: 'ⓒ'

CircledLatinSmallLetterD

\u{24d3}: 'ⓓ'

CircledLatinSmallLetterE

\u{24d4}: 'ⓔ'

CircledLatinSmallLetterF

\u{24d5}: 'ⓕ'

CircledLatinSmallLetterG

\u{24d6}: 'ⓖ'

CircledLatinSmallLetterH

\u{24d7}: 'ⓗ'

CircledLatinSmallLetterI

\u{24d8}: 'ⓘ'

CircledLatinSmallLetterJ

\u{24d9}: 'ⓙ'

CircledLatinSmallLetterK

\u{24da}: 'ⓚ'

CircledLatinSmallLetterL

\u{24db}: 'ⓛ'

CircledLatinSmallLetterM

\u{24dc}: 'ⓜ'

CircledLatinSmallLetterN

\u{24dd}: 'ⓝ'

CircledLatinSmallLetterO

\u{24de}: 'ⓞ'

CircledLatinSmallLetterP

\u{24df}: 'ⓟ'

CircledLatinSmallLetterQ

\u{24e0}: 'ⓠ'

CircledLatinSmallLetterR

\u{24e1}: 'ⓡ'

CircledLatinSmallLetterS

\u{24e2}: 'ⓢ'

CircledLatinSmallLetterT

\u{24e3}: 'ⓣ'

CircledLatinSmallLetterU

\u{24e4}: 'ⓤ'

CircledLatinSmallLetterV

\u{24e5}: 'ⓥ'

CircledLatinSmallLetterW

\u{24e6}: 'ⓦ'

CircledLatinSmallLetterX

\u{24e7}: 'ⓧ'

CircledLatinSmallLetterY

\u{24e8}: 'ⓨ'

CircledLatinSmallLetterZ

\u{24e9}: 'ⓩ'

CircledDigitZero

\u{24ea}: '⓪'

NegativeCircledNumberEleven

\u{24eb}: '⓫'

NegativeCircledNumberTwelve

\u{24ec}: '⓬'

NegativeCircledNumberThirteen

\u{24ed}: '⓭'

NegativeCircledNumberFourteen

\u{24ee}: '⓮'

NegativeCircledNumberFifteen

\u{24ef}: '⓯'

NegativeCircledNumberSixteen

\u{24f0}: '⓰'

NegativeCircledNumberSeventeen

\u{24f1}: '⓱'

NegativeCircledNumberEighteen

\u{24f2}: '⓲'

NegativeCircledNumberNineteen

\u{24f3}: '⓳'

NegativeCircledNumberTwenty

\u{24f4}: '⓴'

DoubleCircledDigitOne

\u{24f5}: '⓵'

DoubleCircledDigitTwo

\u{24f6}: '⓶'

DoubleCircledDigitThree

\u{24f7}: '⓷'

DoubleCircledDigitFour

\u{24f8}: '⓸'

DoubleCircledDigitFive

\u{24f9}: '⓹'

DoubleCircledDigitSix

\u{24fa}: '⓺'

DoubleCircledDigitSeven

\u{24fb}: '⓻'

DoubleCircledDigitEight

\u{24fc}: '⓼'

DoubleCircledDigitNine

\u{24fd}: '⓽'

DoubleCircledNumberTen

\u{24fe}: '⓾'

Methods

impl EnclosedAlphanumerics

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case