Enum unicode_types::enclosed_cjk_letters_and_months::EnclosedCJKLettersandMonths

pub enum EnclosedCJKLettersandMonths {
    ParenthesizedHangulKiyeok,
    ParenthesizedHangulNieun,
    ParenthesizedHangulTikeut,
    ParenthesizedHangulRieul,
    ParenthesizedHangulMieum,
    ParenthesizedHangulPieup,
    ParenthesizedHangulSios,
    ParenthesizedHangulIeung,
    ParenthesizedHangulCieuc,
    ParenthesizedHangulChieuch,
    ParenthesizedHangulKhieukh,
    ParenthesizedHangulThieuth,
    ParenthesizedHangulPhieuph,
    ParenthesizedHangulHieuh,
    ParenthesizedHangulKiyeokA,
    ParenthesizedHangulNieunA,
    ParenthesizedHangulTikeutA,
    ParenthesizedHangulRieulA,
    ParenthesizedHangulMieumA,
    ParenthesizedHangulPieupA,
    ParenthesizedHangulSiosA,
    ParenthesizedHangulIeungA,
    ParenthesizedHangulCieucA,
    ParenthesizedHangulChieuchA,
    ParenthesizedHangulKhieukhA,
    ParenthesizedHangulThieuthA,
    ParenthesizedHangulPhieuphA,
    ParenthesizedHangulHieuhA,
    ParenthesizedHangulCieucU,
    ParenthesizedKoreanCharacterOjeon,
    ParenthesizedKoreanCharacterOHu,
    ParenthesizedIdeographOne,
    ParenthesizedIdeographTwo,
    ParenthesizedIdeographThree,
    ParenthesizedIdeographFour,
    ParenthesizedIdeographFive,
    ParenthesizedIdeographSix,
    ParenthesizedIdeographSeven,
    ParenthesizedIdeographEight,
    ParenthesizedIdeographNine,
    ParenthesizedIdeographTen,
    ParenthesizedIdeographMoon,
    ParenthesizedIdeographFire,
    ParenthesizedIdeographWater,
    ParenthesizedIdeographWood,
    ParenthesizedIdeographMetal,
    ParenthesizedIdeographEarth,
    ParenthesizedIdeographSun,
    ParenthesizedIdeographStock,
    ParenthesizedIdeographHave,
    ParenthesizedIdeographSociety,
    ParenthesizedIdeographName,
    ParenthesizedIdeographSpecial,
    ParenthesizedIdeographFinancial,
    ParenthesizedIdeographCongratulation,
    ParenthesizedIdeographLabor,
    ParenthesizedIdeographRepresent,
    ParenthesizedIdeographCall,
    ParenthesizedIdeographStudy,
    ParenthesizedIdeographSupervise,
    ParenthesizedIdeographEnterprise,
    ParenthesizedIdeographResource,
    ParenthesizedIdeographAlliance,
    ParenthesizedIdeographFestival,
    ParenthesizedIdeographRest,
    ParenthesizedIdeographSelf,
    ParenthesizedIdeographReach,
    CircledIdeographQuestion,
    CircledIdeographKindergarten,
    CircledIdeographSchool,
    CircledIdeographKoto,
    CircledNumberTenOnBlackSquare,
    CircledNumberTwentyOnBlackSquare,
    CircledNumberThirtyOnBlackSquare,
    CircledNumberFortyOnBlackSquare,
    CircledNumberFiftyOnBlackSquare,
    CircledNumberSixtyOnBlackSquare,
    CircledNumberSeventyOnBlackSquare,
    CircledNumberEightyOnBlackSquare,
    PartnershipSign,
    CircledNumberTwentyOne,
    CircledNumberTwentyTwo,
    CircledNumberTwentyThree,
    CircledNumberTwentyFour,
    CircledNumberTwentyFive,
    CircledNumberTwentySix,
    CircledNumberTwentySeven,
    CircledNumberTwentyEight,
    CircledNumberTwentyNine,
    CircledNumberThirty,
    CircledNumberThirtyOne,
    CircledNumberThirtyTwo,
    CircledNumberThirtyThree,
    CircledNumberThirtyFour,
    CircledNumberThirtyFive,
    CircledHangulKiyeok,
    CircledHangulNieun,
    CircledHangulTikeut,
    CircledHangulRieul,
    CircledHangulMieum,
    CircledHangulPieup,
    CircledHangulSios,
    CircledHangulIeung,
    CircledHangulCieuc,
    CircledHangulChieuch,
    CircledHangulKhieukh,
    CircledHangulThieuth,
    CircledHangulPhieuph,
    CircledHangulHieuh,
    CircledHangulKiyeokA,
    CircledHangulNieunA,
    CircledHangulTikeutA,
    CircledHangulRieulA,
    CircledHangulMieumA,
    CircledHangulPieupA,
    CircledHangulSiosA,
    CircledHangulIeungA,
    CircledHangulCieucA,
    CircledHangulChieuchA,
    CircledHangulKhieukhA,
    CircledHangulThieuthA,
    CircledHangulPhieuphA,
    CircledHangulHieuhA,
    CircledKoreanCharacterChamko,
    CircledKoreanCharacterJueui,
    CircledHangulIeungU,
    KoreanStandardSymbol,
    CircledIdeographOne,
    CircledIdeographTwo,
    CircledIdeographThree,
    CircledIdeographFour,
    CircledIdeographFive,
    CircledIdeographSix,
    CircledIdeographSeven,
    CircledIdeographEight,
    CircledIdeographNine,
    CircledIdeographTen,
    CircledIdeographMoon,
    CircledIdeographFire,
    CircledIdeographWater,
    CircledIdeographWood,
    CircledIdeographMetal,
    CircledIdeographEarth,
    CircledIdeographSun,
    CircledIdeographStock,
    CircledIdeographHave,
    CircledIdeographSociety,
    CircledIdeographName,
    CircledIdeographSpecial,
    CircledIdeographFinancial,
    CircledIdeographCongratulation,
    CircledIdeographLabor,
    CircledIdeographSecret,
    CircledIdeographMale,
    CircledIdeographFemale,
    CircledIdeographSuitable,
    CircledIdeographExcellent,
    CircledIdeographPrint,
    CircledIdeographAttention,
    CircledIdeographItem,
    CircledIdeographRest,
    CircledIdeographCopy,
    CircledIdeographCorrect,
    CircledIdeographHigh,
    CircledIdeographCentre,
    CircledIdeographLow,
    CircledIdeographLeft,
    CircledIdeographRight,
    CircledIdeographMedicine,
    CircledIdeographReligion,
    CircledIdeographStudy,
    CircledIdeographSupervise,
    CircledIdeographEnterprise,
    CircledIdeographResource,
    CircledIdeographAlliance,
    CircledIdeographNight,
    CircledNumberThirtySix,
    CircledNumberThirtySeven,
    CircledNumberThirtyEight,
    CircledNumberThirtyNine,
    CircledNumberForty,
    CircledNumberFortyOne,
    CircledNumberFortyTwo,
    CircledNumberFortyThree,
    CircledNumberFortyFour,
    CircledNumberFortyFive,
    CircledNumberFortySix,
    CircledNumberFortySeven,
    CircledNumberFortyEight,
    CircledNumberFortyNine,
    CircledNumberFifty,
    IdeographicTelegraphSymbolForJanuary,
    IdeographicTelegraphSymbolForFebruary,
    IdeographicTelegraphSymbolForMarch,
    IdeographicTelegraphSymbolForApril,
    IdeographicTelegraphSymbolForMay,
    IdeographicTelegraphSymbolForJune,
    IdeographicTelegraphSymbolForJuly,
    IdeographicTelegraphSymbolForAugust,
    IdeographicTelegraphSymbolForSeptember,
    IdeographicTelegraphSymbolForOctober,
    IdeographicTelegraphSymbolForNovember,
    IdeographicTelegraphSymbolForDecember,
    SquareHg,
    SquareErg,
    SquareEv,
    LimitedLiabilitySign,
    CircledKatakanaA,
    CircledKatakanaI,
    CircledKatakanaU,
    CircledKatakanaE,
    CircledKatakanaO,
    CircledKatakanaKa,
    CircledKatakanaKi,
    CircledKatakanaKu,
    CircledKatakanaKe,
    CircledKatakanaKo,
    CircledKatakanaSa,
    CircledKatakanaSi,
    CircledKatakanaSu,
    CircledKatakanaSe,
    CircledKatakanaSo,
    CircledKatakanaTa,
    CircledKatakanaTi,
    CircledKatakanaTu,
    CircledKatakanaTe,
    CircledKatakanaTo,
    CircledKatakanaNa,
    CircledKatakanaNi,
    CircledKatakanaNu,
    CircledKatakanaNe,
    CircledKatakanaNo,
    CircledKatakanaHa,
    CircledKatakanaHi,
    CircledKatakanaHu,
    CircledKatakanaHe,
    CircledKatakanaHo,
    CircledKatakanaMa,
    CircledKatakanaMi,
    CircledKatakanaMu,
    CircledKatakanaMe,
    CircledKatakanaMo,
    CircledKatakanaYa,
    CircledKatakanaYu,
    CircledKatakanaYo,
    CircledKatakanaRa,
    CircledKatakanaRi,
    CircledKatakanaRu,
    CircledKatakanaRe,
    CircledKatakanaRo,
    CircledKatakanaWa,
    CircledKatakanaWi,
    CircledKatakanaWe,
    CircledKatakanaWo,
}

An enum to represent all characters in the EnclosedCJKLettersandMonths block.

Variants
ParenthesizedHangulKiyeok

\u{3200}: '㈀'

ParenthesizedHangulNieun

\u{3201}: '㈁'

ParenthesizedHangulTikeut

\u{3202}: '㈂'

ParenthesizedHangulRieul

\u{3203}: '㈃'

ParenthesizedHangulMieum

\u{3204}: '㈄'

ParenthesizedHangulPieup

\u{3205}: '㈅'

ParenthesizedHangulSios

\u{3206}: '㈆'

ParenthesizedHangulIeung

\u{3207}: '㈇'

ParenthesizedHangulCieuc

\u{3208}: '㈈'

ParenthesizedHangulChieuch

\u{3209}: '㈉'

ParenthesizedHangulKhieukh

\u{320a}: '㈊'

ParenthesizedHangulThieuth

\u{320b}: '㈋'

ParenthesizedHangulPhieuph

\u{320c}: '㈌'

ParenthesizedHangulHieuh

\u{320d}: '㈍'

ParenthesizedHangulKiyeokA

\u{320e}: '㈎'

ParenthesizedHangulNieunA

\u{320f}: '㈏'

ParenthesizedHangulTikeutA

\u{3210}: '㈐'

ParenthesizedHangulRieulA

\u{3211}: '㈑'

ParenthesizedHangulMieumA

\u{3212}: '㈒'

ParenthesizedHangulPieupA

\u{3213}: '㈓'

ParenthesizedHangulSiosA

\u{3214}: '㈔'

ParenthesizedHangulIeungA

\u{3215}: '㈕'

ParenthesizedHangulCieucA

\u{3216}: '㈖'

ParenthesizedHangulChieuchA

\u{3217}: '㈗'

ParenthesizedHangulKhieukhA

\u{3218}: '㈘'

ParenthesizedHangulThieuthA

\u{3219}: '㈙'

ParenthesizedHangulPhieuphA

\u{321a}: '㈚'

ParenthesizedHangulHieuhA

\u{321b}: '㈛'

ParenthesizedHangulCieucU

\u{321c}: '㈜'

ParenthesizedKoreanCharacterOjeon

\u{321d}: '㈝'

ParenthesizedKoreanCharacterOHu

\u{321e}: '㈞'

ParenthesizedIdeographOne

\u{3220}: '㈠'

ParenthesizedIdeographTwo

\u{3221}: '㈡'

ParenthesizedIdeographThree

\u{3222}: '㈢'

ParenthesizedIdeographFour

\u{3223}: '㈣'

ParenthesizedIdeographFive

\u{3224}: '㈤'

ParenthesizedIdeographSix

\u{3225}: '㈥'

ParenthesizedIdeographSeven

\u{3226}: '㈦'

ParenthesizedIdeographEight

\u{3227}: '㈧'

ParenthesizedIdeographNine

\u{3228}: '㈨'

ParenthesizedIdeographTen

\u{3229}: '㈩'

ParenthesizedIdeographMoon

\u{322a}: '㈪'

ParenthesizedIdeographFire

\u{322b}: '㈫'

ParenthesizedIdeographWater

\u{322c}: '㈬'

ParenthesizedIdeographWood

\u{322d}: '㈭'

ParenthesizedIdeographMetal

\u{322e}: '㈮'

ParenthesizedIdeographEarth

\u{322f}: '㈯'

ParenthesizedIdeographSun

\u{3230}: '㈰'

ParenthesizedIdeographStock

\u{3231}: '㈱'

ParenthesizedIdeographHave

\u{3232}: '㈲'

ParenthesizedIdeographSociety

\u{3233}: '㈳'

ParenthesizedIdeographName

\u{3234}: '㈴'

ParenthesizedIdeographSpecial

\u{3235}: '㈵'

ParenthesizedIdeographFinancial

\u{3236}: '㈶'

ParenthesizedIdeographCongratulation

\u{3237}: '㈷'

ParenthesizedIdeographLabor

\u{3238}: '㈸'

ParenthesizedIdeographRepresent

\u{3239}: '㈹'

ParenthesizedIdeographCall

\u{323a}: '㈺'

ParenthesizedIdeographStudy

\u{323b}: '㈻'

ParenthesizedIdeographSupervise

\u{323c}: '㈼'

ParenthesizedIdeographEnterprise

\u{323d}: '㈽'

ParenthesizedIdeographResource

\u{323e}: '㈾'

ParenthesizedIdeographAlliance

\u{323f}: '㈿'

ParenthesizedIdeographFestival

\u{3240}: '㉀'

ParenthesizedIdeographRest

\u{3241}: '㉁'

ParenthesizedIdeographSelf

\u{3242}: '㉂'

ParenthesizedIdeographReach

\u{3243}: '㉃'

CircledIdeographQuestion

\u{3244}: '㉄'

CircledIdeographKindergarten

\u{3245}: '㉅'

CircledIdeographSchool

\u{3246}: '㉆'

CircledIdeographKoto

\u{3247}: '㉇'

CircledNumberTenOnBlackSquare

\u{3248}: '㉈'

CircledNumberTwentyOnBlackSquare

\u{3249}: '㉉'

CircledNumberThirtyOnBlackSquare

\u{324a}: '㉊'

CircledNumberFortyOnBlackSquare

\u{324b}: '㉋'

CircledNumberFiftyOnBlackSquare

\u{324c}: '㉌'

CircledNumberSixtyOnBlackSquare

\u{324d}: '㉍'

CircledNumberSeventyOnBlackSquare

\u{324e}: '㉎'

CircledNumberEightyOnBlackSquare

\u{324f}: '㉏'

PartnershipSign

\u{3250}: '㉐'

CircledNumberTwentyOne

\u{3251}: '㉑'

CircledNumberTwentyTwo

\u{3252}: '㉒'

CircledNumberTwentyThree

\u{3253}: '㉓'

CircledNumberTwentyFour

\u{3254}: '㉔'

CircledNumberTwentyFive

\u{3255}: '㉕'

CircledNumberTwentySix

\u{3256}: '㉖'

CircledNumberTwentySeven

\u{3257}: '㉗'

CircledNumberTwentyEight

\u{3258}: '㉘'

CircledNumberTwentyNine

\u{3259}: '㉙'

CircledNumberThirty

\u{325a}: '㉚'

CircledNumberThirtyOne

\u{325b}: '㉛'

CircledNumberThirtyTwo

\u{325c}: '㉜'

CircledNumberThirtyThree

\u{325d}: '㉝'

CircledNumberThirtyFour

\u{325e}: '㉞'

CircledNumberThirtyFive

\u{325f}: '㉟'

CircledHangulKiyeok

\u{3260}: '㉠'

CircledHangulNieun

\u{3261}: '㉡'

CircledHangulTikeut

\u{3262}: '㉢'

CircledHangulRieul

\u{3263}: '㉣'

CircledHangulMieum

\u{3264}: '㉤'

CircledHangulPieup

\u{3265}: '㉥'

CircledHangulSios

\u{3266}: '㉦'

CircledHangulIeung

\u{3267}: '㉧'

CircledHangulCieuc

\u{3268}: '㉨'

CircledHangulChieuch

\u{3269}: '㉩'

CircledHangulKhieukh

\u{326a}: '㉪'

CircledHangulThieuth

\u{326b}: '㉫'

CircledHangulPhieuph

\u{326c}: '㉬'

CircledHangulHieuh

\u{326d}: '㉭'

CircledHangulKiyeokA

\u{326e}: '㉮'

CircledHangulNieunA

\u{326f}: '㉯'

CircledHangulTikeutA

\u{3270}: '㉰'

CircledHangulRieulA

\u{3271}: '㉱'

CircledHangulMieumA

\u{3272}: '㉲'

CircledHangulPieupA

\u{3273}: '㉳'

CircledHangulSiosA

\u{3274}: '㉴'

CircledHangulIeungA

\u{3275}: '㉵'

CircledHangulCieucA

\u{3276}: '㉶'

CircledHangulChieuchA

\u{3277}: '㉷'

CircledHangulKhieukhA

\u{3278}: '㉸'

CircledHangulThieuthA

\u{3279}: '㉹'

CircledHangulPhieuphA

\u{327a}: '㉺'

CircledHangulHieuhA

\u{327b}: '㉻'

CircledKoreanCharacterChamko

\u{327c}: '㉼'

CircledKoreanCharacterJueui

\u{327d}: '㉽'

CircledHangulIeungU

\u{327e}: '㉾'

KoreanStandardSymbol

\u{327f}: '㉿'

CircledIdeographOne

\u{3280}: '㊀'

CircledIdeographTwo

\u{3281}: '㊁'

CircledIdeographThree

\u{3282}: '㊂'

CircledIdeographFour

\u{3283}: '㊃'

CircledIdeographFive

\u{3284}: '㊄'

CircledIdeographSix

\u{3285}: '㊅'

CircledIdeographSeven

\u{3286}: '㊆'

CircledIdeographEight

\u{3287}: '㊇'

CircledIdeographNine

\u{3288}: '㊈'

CircledIdeographTen

\u{3289}: '㊉'

CircledIdeographMoon

\u{328a}: '㊊'

CircledIdeographFire

\u{328b}: '㊋'

CircledIdeographWater

\u{328c}: '㊌'

CircledIdeographWood

\u{328d}: '㊍'

CircledIdeographMetal

\u{328e}: '㊎'

CircledIdeographEarth

\u{328f}: '㊏'

CircledIdeographSun

\u{3290}: '㊐'

CircledIdeographStock

\u{3291}: '㊑'

CircledIdeographHave

\u{3292}: '㊒'

CircledIdeographSociety

\u{3293}: '㊓'

CircledIdeographName

\u{3294}: '㊔'

CircledIdeographSpecial

\u{3295}: '㊕'

CircledIdeographFinancial

\u{3296}: '㊖'

CircledIdeographCongratulation

\u{3297}: '㊗'

CircledIdeographLabor

\u{3298}: '㊘'

CircledIdeographSecret

\u{3299}: '㊙'

CircledIdeographMale

\u{329a}: '㊚'

CircledIdeographFemale

\u{329b}: '㊛'

CircledIdeographSuitable

\u{329c}: '㊜'

CircledIdeographExcellent

\u{329d}: '㊝'

CircledIdeographPrint

\u{329e}: '㊞'

CircledIdeographAttention

\u{329f}: '㊟'

CircledIdeographItem

\u{32a0}: '㊠'

CircledIdeographRest

\u{32a1}: '㊡'

CircledIdeographCopy

\u{32a2}: '㊢'

CircledIdeographCorrect

\u{32a3}: '㊣'

CircledIdeographHigh

\u{32a4}: '㊤'

CircledIdeographCentre

\u{32a5}: '㊥'

CircledIdeographLow

\u{32a6}: '㊦'

CircledIdeographLeft

\u{32a7}: '㊧'

CircledIdeographRight

\u{32a8}: '㊨'

CircledIdeographMedicine

\u{32a9}: '㊩'

CircledIdeographReligion

\u{32aa}: '㊪'

CircledIdeographStudy

\u{32ab}: '㊫'

CircledIdeographSupervise

\u{32ac}: '㊬'

CircledIdeographEnterprise

\u{32ad}: '㊭'

CircledIdeographResource

\u{32ae}: '㊮'

CircledIdeographAlliance

\u{32af}: '㊯'

CircledIdeographNight

\u{32b0}: '㊰'

CircledNumberThirtySix

\u{32b1}: '㊱'

CircledNumberThirtySeven

\u{32b2}: '㊲'

CircledNumberThirtyEight

\u{32b3}: '㊳'

CircledNumberThirtyNine

\u{32b4}: '㊴'

CircledNumberForty

\u{32b5}: '㊵'

CircledNumberFortyOne

\u{32b6}: '㊶'

CircledNumberFortyTwo

\u{32b7}: '㊷'

CircledNumberFortyThree

\u{32b8}: '㊸'

CircledNumberFortyFour

\u{32b9}: '㊹'

CircledNumberFortyFive

\u{32ba}: '㊺'

CircledNumberFortySix

\u{32bb}: '㊻'

CircledNumberFortySeven

\u{32bc}: '㊼'

CircledNumberFortyEight

\u{32bd}: '㊽'

CircledNumberFortyNine

\u{32be}: '㊾'

CircledNumberFifty

\u{32bf}: '㊿'

IdeographicTelegraphSymbolForJanuary

\u{32c0}: '㋀'

IdeographicTelegraphSymbolForFebruary

\u{32c1}: '㋁'

IdeographicTelegraphSymbolForMarch

\u{32c2}: '㋂'

IdeographicTelegraphSymbolForApril

\u{32c3}: '㋃'

IdeographicTelegraphSymbolForMay

\u{32c4}: '㋄'

IdeographicTelegraphSymbolForJune

\u{32c5}: '㋅'

IdeographicTelegraphSymbolForJuly

\u{32c6}: '㋆'

IdeographicTelegraphSymbolForAugust

\u{32c7}: '㋇'

IdeographicTelegraphSymbolForSeptember

\u{32c8}: '㋈'

IdeographicTelegraphSymbolForOctober

\u{32c9}: '㋉'

IdeographicTelegraphSymbolForNovember

\u{32ca}: '㋊'

IdeographicTelegraphSymbolForDecember

\u{32cb}: '㋋'

SquareHg

\u{32cc}: '㋌'

SquareErg

\u{32cd}: '㋍'

SquareEv

\u{32ce}: '㋎'

LimitedLiabilitySign

\u{32cf}: '㋏'

CircledKatakanaA

\u{32d0}: '㋐'

CircledKatakanaI

\u{32d1}: '㋑'

CircledKatakanaU

\u{32d2}: '㋒'

CircledKatakanaE

\u{32d3}: '㋓'

CircledKatakanaO

\u{32d4}: '㋔'

CircledKatakanaKa

\u{32d5}: '㋕'

CircledKatakanaKi

\u{32d6}: '㋖'

CircledKatakanaKu

\u{32d7}: '㋗'

CircledKatakanaKe

\u{32d8}: '㋘'

CircledKatakanaKo

\u{32d9}: '㋙'

CircledKatakanaSa

\u{32da}: '㋚'

CircledKatakanaSi

\u{32db}: '㋛'

CircledKatakanaSu

\u{32dc}: '㋜'

CircledKatakanaSe

\u{32dd}: '㋝'

CircledKatakanaSo

\u{32de}: '㋞'

CircledKatakanaTa

\u{32df}: '㋟'

CircledKatakanaTi

\u{32e0}: '㋠'

CircledKatakanaTu

\u{32e1}: '㋡'

CircledKatakanaTe

\u{32e2}: '㋢'

CircledKatakanaTo

\u{32e3}: '㋣'

CircledKatakanaNa

\u{32e4}: '㋤'

CircledKatakanaNi

\u{32e5}: '㋥'

CircledKatakanaNu

\u{32e6}: '㋦'

CircledKatakanaNe

\u{32e7}: '㋧'

CircledKatakanaNo

\u{32e8}: '㋨'

CircledKatakanaHa

\u{32e9}: '㋩'

CircledKatakanaHi

\u{32ea}: '㋪'

CircledKatakanaHu

\u{32eb}: '㋫'

CircledKatakanaHe

\u{32ec}: '㋬'

CircledKatakanaHo

\u{32ed}: '㋭'

CircledKatakanaMa

\u{32ee}: '㋮'

CircledKatakanaMi

\u{32ef}: '㋯'

CircledKatakanaMu

\u{32f0}: '㋰'

CircledKatakanaMe

\u{32f1}: '㋱'

CircledKatakanaMo

\u{32f2}: '㋲'

CircledKatakanaYa

\u{32f3}: '㋳'

CircledKatakanaYu

\u{32f4}: '㋴'

CircledKatakanaYo

\u{32f5}: '㋵'

CircledKatakanaRa

\u{32f6}: '㋶'

CircledKatakanaRi

\u{32f7}: '㋷'

CircledKatakanaRu

\u{32f8}: '㋸'

CircledKatakanaRe

\u{32f9}: '㋹'

CircledKatakanaRo

\u{32fa}: '㋺'

CircledKatakanaWa

\u{32fb}: '㋻'

CircledKatakanaWi

\u{32fc}: '㋼'

CircledKatakanaWe

\u{32fd}: '㋽'

CircledKatakanaWo

\u{32fe}: '㋾'

Methods

impl EnclosedCJKLettersandMonths

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case