Enum unicode_types::ethiopic::Ethiopic

pub enum Ethiopic {
    SyllableHa,
    SyllableHu,
    SyllableHi,
    SyllableHaa,
    SyllableHee,
    SyllableHe,
    SyllableHo,
    SyllableHoa,
    SyllableLa,
    SyllableLu,
    SyllableLi,
    SyllableLaa,
    SyllableLee,
    SyllableLe,
    SyllableLo,
    SyllableLwa,
    SyllableHha,
    SyllableHhu,
    SyllableHhi,
    SyllableHhaa,
    SyllableHhee,
    SyllableHhe,
    SyllableHho,
    SyllableHhwa,
    SyllableMa,
    SyllableMu,
    SyllableMi,
    SyllableMaa,
    SyllableMee,
    SyllableMe,
    SyllableMo,
    SyllableMwa,
    SyllableSza,
    SyllableSzu,
    SyllableSzi,
    SyllableSzaa,
    SyllableSzee,
    SyllableSze,
    SyllableSzo,
    SyllableSzwa,
    SyllableRa,
    SyllableRu,
    SyllableRi,
    SyllableRaa,
    SyllableRee,
    SyllableRe,
    SyllableRo,
    SyllableRwa,
    SyllableSa,
    SyllableSu,
    SyllableSi,
    SyllableSaa,
    SyllableSee,
    SyllableSe,
    SyllableSo,
    SyllableSwa,
    SyllableSha,
    SyllableShu,
    SyllableShi,
    SyllableShaa,
    SyllableShee,
    SyllableShe,
    SyllableSho,
    SyllableShwa,
    SyllableQa,
    SyllableQu,
    SyllableQi,
    SyllableQaa,
    SyllableQee,
    SyllableQe,
    SyllableQo,
    SyllableQoa,
    SyllableQwa,
    SyllableQwi,
    SyllableQwaa,
    SyllableQwee,
    SyllableQwe,
    SyllableQha,
    SyllableQhu,
    SyllableQhi,
    SyllableQhaa,
    SyllableQhee,
    SyllableQhe,
    SyllableQho,
    SyllableQhwa,
    SyllableQhwi,
    SyllableQhwaa,
    SyllableQhwee,
    SyllableQhwe,
    SyllableBa,
    SyllableBu,
    SyllableBi,
    SyllableBaa,
    SyllableBee,
    SyllableBe,
    SyllableBo,
    SyllableBwa,
    SyllableVa,
    SyllableVu,
    SyllableVi,
    SyllableVaa,
    SyllableVee,
    SyllableVe,
    SyllableVo,
    SyllableVwa,
    SyllableTa,
    SyllableTu,
    SyllableTi,
    SyllableTaa,
    SyllableTee,
    SyllableTe,
    SyllableTo,
    SyllableTwa,
    SyllableCa,
    SyllableCu,
    SyllableCi,
    SyllableCaa,
    SyllableCee,
    SyllableCe,
    SyllableCo,
    SyllableCwa,
    SyllableXa,
    SyllableXu,
    SyllableXi,
    SyllableXaa,
    SyllableXee,
    SyllableXe,
    SyllableXo,
    SyllableXoa,
    SyllableXwa,
    SyllableXwi,
    SyllableXwaa,
    SyllableXwee,
    SyllableXwe,
    SyllableNa,
    SyllableNu,
    SyllableNi,
    SyllableNaa,
    SyllableNee,
    SyllableNe,
    SyllableNo,
    SyllableNwa,
    SyllableNya,
    SyllableNyu,
    SyllableNyi,
    SyllableNyaa,
    SyllableNyee,
    SyllableNye,
    SyllableNyo,
    SyllableNywa,
    SyllableGlottalA,
    SyllableGlottalU,
    SyllableGlottalI,
    SyllableGlottalAa,
    SyllableGlottalEe,
    SyllableGlottalE,
    SyllableGlottalO,
    SyllableGlottalWa,
    SyllableKa,
    SyllableKu,
    SyllableKi,
    SyllableKaa,
    SyllableKee,
    SyllableKe,
    SyllableKo,
    SyllableKoa,
    SyllableKwa,
    SyllableKwi,
    SyllableKwaa,
    SyllableKwee,
    SyllableKwe,
    SyllableKxa,
    SyllableKxu,
    SyllableKxi,
    SyllableKxaa,
    SyllableKxee,
    SyllableKxe,
    SyllableKxo,
    SyllableKxwa,
    SyllableKxwi,
    SyllableKxwaa,
    SyllableKxwee,
    SyllableKxwe,
    SyllableWa,
    SyllableWu,
    SyllableWi,
    SyllableWaa,
    SyllableWee,
    SyllableWe,
    SyllableWo,
    SyllableWoa,
    SyllablePharyngealA,
    SyllablePharyngealU,
    SyllablePharyngealI,
    SyllablePharyngealAa,
    SyllablePharyngealEe,
    SyllablePharyngealE,
    SyllablePharyngealO,
    SyllableZa,
    SyllableZu,
    SyllableZi,
    SyllableZaa,
    SyllableZee,
    SyllableZe,
    SyllableZo,
    SyllableZwa,
    SyllableZha,
    SyllableZhu,
    SyllableZhi,
    SyllableZhaa,
    SyllableZhee,
    SyllableZhe,
    SyllableZho,
    SyllableZhwa,
    SyllableYa,
    SyllableYu,
    SyllableYi,
    SyllableYaa,
    SyllableYee,
    SyllableYe,
    SyllableYo,
    SyllableYoa,
    SyllableDa,
    SyllableDu,
    SyllableDi,
    SyllableDaa,
    SyllableDee,
    SyllableDe,
    SyllableDo,
    SyllableDwa,
    SyllableDda,
    SyllableDdu,
    SyllableDdi,
    SyllableDdaa,
    SyllableDdee,
    SyllableDde,
    SyllableDdo,
    SyllableDdwa,
    SyllableJa,
    SyllableJu,
    SyllableJi,
    SyllableJaa,
    SyllableJee,
    SyllableJe,
    SyllableJo,
    SyllableJwa,
    SyllableGa,
    SyllableGu,
    SyllableGi,
    SyllableGaa,
    SyllableGee,
    SyllableGe,
    SyllableGo,
    SyllableGoa,
    SyllableGwa,
    SyllableGwi,
    SyllableGwaa,
    SyllableGwee,
    SyllableGwe,
    SyllableGga,
    SyllableGgu,
    SyllableGgi,
    SyllableGgaa,
    SyllableGgee,
    SyllableGge,
    SyllableGgo,
    SyllableGgwaa,
    SyllableTha,
    SyllableThu,
    SyllableThi,
    SyllableThaa,
    SyllableThee,
    SyllableThe,
    SyllableTho,
    SyllableThwa,
    SyllableCha,
    SyllableChu,
    SyllableChi,
    SyllableChaa,
    SyllableChee,
    SyllableChe,
    SyllableCho,
    SyllableChwa,
    SyllablePha,
    SyllablePhu,
    SyllablePhi,
    SyllablePhaa,
    SyllablePhee,
    SyllablePhe,
    SyllablePho,
    SyllablePhwa,
    SyllableTsa,
    SyllableTsu,
    SyllableTsi,
    SyllableTsaa,
    SyllableTsee,
    SyllableTse,
    SyllableTso,
    SyllableTswa,
    SyllableTza,
    SyllableTzu,
    SyllableTzi,
    SyllableTzaa,
    SyllableTzee,
    SyllableTze,
    SyllableTzo,
    SyllableTzoa,
    SyllableFa,
    SyllableFu,
    SyllableFi,
    SyllableFaa,
    SyllableFee,
    SyllableFe,
    SyllableFo,
    SyllableFwa,
    SyllablePa,
    SyllablePu,
    SyllablePi,
    SyllablePaa,
    SyllablePee,
    SyllablePe,
    SyllablePo,
    SyllablePwa,
    SyllableRya,
    SyllableMya,
    SyllableFya,
    CombiningGeminationAndVowelLengthMark,
    CombiningVowelLengthMark,
    CombiningGeminationMark,
    SectionMark,
    Wordspace,
    FullStop,
    Comma,
    Semicolon,
    Colon,
    PrefaceColon,
    QuestionMark,
    ParagraphSeparator,
    DigitOne,
    DigitTwo,
    DigitThree,
    DigitFour,
    DigitFive,
    DigitSix,
    DigitSeven,
    DigitEight,
    DigitNine,
    NumberTen,
    NumberTwenty,
    NumberThirty,
    NumberForty,
    NumberFifty,
    NumberSixty,
    NumberSeventy,
    NumberEighty,
    NumberNinety,
    NumberHundred,
    NumberTenThousand,
}

An enum to represent all characters in the Ethiopic block.

Variants
SyllableHa

\u{1200}: 'ሀ'

SyllableHu

\u{1201}: 'ሁ'

SyllableHi

\u{1202}: 'ሂ'

SyllableHaa

\u{1203}: 'ሃ'

SyllableHee

\u{1204}: 'ሄ'

SyllableHe

\u{1205}: 'ህ'

SyllableHo

\u{1206}: 'ሆ'

SyllableHoa

\u{1207}: 'ሇ'

SyllableLa

\u{1208}: 'ለ'

SyllableLu

\u{1209}: 'ሉ'

SyllableLi

\u{120a}: 'ሊ'

SyllableLaa

\u{120b}: 'ላ'

SyllableLee

\u{120c}: 'ሌ'

SyllableLe

\u{120d}: 'ል'

SyllableLo

\u{120e}: 'ሎ'

SyllableLwa

\u{120f}: 'ሏ'

SyllableHha

\u{1210}: 'ሐ'

SyllableHhu

\u{1211}: 'ሑ'

SyllableHhi

\u{1212}: 'ሒ'

SyllableHhaa

\u{1213}: 'ሓ'

SyllableHhee

\u{1214}: 'ሔ'

SyllableHhe

\u{1215}: 'ሕ'

SyllableHho

\u{1216}: 'ሖ'

SyllableHhwa

\u{1217}: 'ሗ'

SyllableMa

\u{1218}: 'መ'

SyllableMu

\u{1219}: 'ሙ'

SyllableMi

\u{121a}: 'ሚ'

SyllableMaa

\u{121b}: 'ማ'

SyllableMee

\u{121c}: 'ሜ'

SyllableMe

\u{121d}: 'ም'

SyllableMo

\u{121e}: 'ሞ'

SyllableMwa

\u{121f}: 'ሟ'

SyllableSza

\u{1220}: 'ሠ'

SyllableSzu

\u{1221}: 'ሡ'

SyllableSzi

\u{1222}: 'ሢ'

SyllableSzaa

\u{1223}: 'ሣ'

SyllableSzee

\u{1224}: 'ሤ'

SyllableSze

\u{1225}: 'ሥ'

SyllableSzo

\u{1226}: 'ሦ'

SyllableSzwa

\u{1227}: 'ሧ'

SyllableRa

\u{1228}: 'ረ'

SyllableRu

\u{1229}: 'ሩ'

SyllableRi

\u{122a}: 'ሪ'

SyllableRaa

\u{122b}: 'ራ'

SyllableRee

\u{122c}: 'ሬ'

SyllableRe

\u{122d}: 'ር'

SyllableRo

\u{122e}: 'ሮ'

SyllableRwa

\u{122f}: 'ሯ'

SyllableSa

\u{1230}: 'ሰ'

SyllableSu

\u{1231}: 'ሱ'

SyllableSi

\u{1232}: 'ሲ'

SyllableSaa

\u{1233}: 'ሳ'

SyllableSee

\u{1234}: 'ሴ'

SyllableSe

\u{1235}: 'ስ'

SyllableSo

\u{1236}: 'ሶ'

SyllableSwa

\u{1237}: 'ሷ'

SyllableSha

\u{1238}: 'ሸ'

SyllableShu

\u{1239}: 'ሹ'

SyllableShi

\u{123a}: 'ሺ'

SyllableShaa

\u{123b}: 'ሻ'

SyllableShee

\u{123c}: 'ሼ'

SyllableShe

\u{123d}: 'ሽ'

SyllableSho

\u{123e}: 'ሾ'

SyllableShwa

\u{123f}: 'ሿ'

SyllableQa

\u{1240}: 'ቀ'

SyllableQu

\u{1241}: 'ቁ'

SyllableQi

\u{1242}: 'ቂ'

SyllableQaa

\u{1243}: 'ቃ'

SyllableQee

\u{1244}: 'ቄ'

SyllableQe

\u{1245}: 'ቅ'

SyllableQo

\u{1246}: 'ቆ'

SyllableQoa

\u{1247}: 'ቇ'

SyllableQwa

\u{1248}: 'ቈ'

SyllableQwi

\u{124a}: 'ቊ'

SyllableQwaa

\u{124b}: 'ቋ'

SyllableQwee

\u{124c}: 'ቌ'

SyllableQwe

\u{124d}: 'ቍ'

SyllableQha

\u{1250}: 'ቐ'

SyllableQhu

\u{1251}: 'ቑ'

SyllableQhi

\u{1252}: 'ቒ'

SyllableQhaa

\u{1253}: 'ቓ'

SyllableQhee

\u{1254}: 'ቔ'

SyllableQhe

\u{1255}: 'ቕ'

SyllableQho

\u{1256}: 'ቖ'

SyllableQhwa

\u{1258}: 'ቘ'

SyllableQhwi

\u{125a}: 'ቚ'

SyllableQhwaa

\u{125b}: 'ቛ'

SyllableQhwee

\u{125c}: 'ቜ'

SyllableQhwe

\u{125d}: 'ቝ'

SyllableBa

\u{1260}: 'በ'

SyllableBu

\u{1261}: 'ቡ'

SyllableBi

\u{1262}: 'ቢ'

SyllableBaa

\u{1263}: 'ባ'

SyllableBee

\u{1264}: 'ቤ'

SyllableBe

\u{1265}: 'ብ'

SyllableBo

\u{1266}: 'ቦ'

SyllableBwa

\u{1267}: 'ቧ'

SyllableVa

\u{1268}: 'ቨ'

SyllableVu

\u{1269}: 'ቩ'

SyllableVi

\u{126a}: 'ቪ'

SyllableVaa

\u{126b}: 'ቫ'

SyllableVee

\u{126c}: 'ቬ'

SyllableVe

\u{126d}: 'ቭ'

SyllableVo

\u{126e}: 'ቮ'

SyllableVwa

\u{126f}: 'ቯ'

SyllableTa

\u{1270}: 'ተ'

SyllableTu

\u{1271}: 'ቱ'

SyllableTi

\u{1272}: 'ቲ'

SyllableTaa

\u{1273}: 'ታ'

SyllableTee

\u{1274}: 'ቴ'

SyllableTe

\u{1275}: 'ት'

SyllableTo

\u{1276}: 'ቶ'

SyllableTwa

\u{1277}: 'ቷ'

SyllableCa

\u{1278}: 'ቸ'

SyllableCu

\u{1279}: 'ቹ'

SyllableCi

\u{127a}: 'ቺ'

SyllableCaa

\u{127b}: 'ቻ'

SyllableCee

\u{127c}: 'ቼ'

SyllableCe

\u{127d}: 'ች'

SyllableCo

\u{127e}: 'ቾ'

SyllableCwa

\u{127f}: 'ቿ'

SyllableXa

\u{1280}: 'ኀ'

SyllableXu

\u{1281}: 'ኁ'

SyllableXi

\u{1282}: 'ኂ'

SyllableXaa

\u{1283}: 'ኃ'

SyllableXee

\u{1284}: 'ኄ'

SyllableXe

\u{1285}: 'ኅ'

SyllableXo

\u{1286}: 'ኆ'

SyllableXoa

\u{1287}: 'ኇ'

SyllableXwa

\u{1288}: 'ኈ'

SyllableXwi

\u{128a}: 'ኊ'

SyllableXwaa

\u{128b}: 'ኋ'

SyllableXwee

\u{128c}: 'ኌ'

SyllableXwe

\u{128d}: 'ኍ'

SyllableNa

\u{1290}: 'ነ'

SyllableNu

\u{1291}: 'ኑ'

SyllableNi

\u{1292}: 'ኒ'

SyllableNaa

\u{1293}: 'ና'

SyllableNee

\u{1294}: 'ኔ'

SyllableNe

\u{1295}: 'ን'

SyllableNo

\u{1296}: 'ኖ'

SyllableNwa

\u{1297}: 'ኗ'

SyllableNya

\u{1298}: 'ኘ'

SyllableNyu

\u{1299}: 'ኙ'

SyllableNyi

\u{129a}: 'ኚ'

SyllableNyaa

\u{129b}: 'ኛ'

SyllableNyee

\u{129c}: 'ኜ'

SyllableNye

\u{129d}: 'ኝ'

SyllableNyo

\u{129e}: 'ኞ'

SyllableNywa

\u{129f}: 'ኟ'

SyllableGlottalA

\u{12a0}: 'አ'

SyllableGlottalU

\u{12a1}: 'ኡ'

SyllableGlottalI

\u{12a2}: 'ኢ'

SyllableGlottalAa

\u{12a3}: 'ኣ'

SyllableGlottalEe

\u{12a4}: 'ኤ'

SyllableGlottalE

\u{12a5}: 'እ'

SyllableGlottalO

\u{12a6}: 'ኦ'

SyllableGlottalWa

\u{12a7}: 'ኧ'

SyllableKa

\u{12a8}: 'ከ'

SyllableKu

\u{12a9}: 'ኩ'

SyllableKi

\u{12aa}: 'ኪ'

SyllableKaa

\u{12ab}: 'ካ'

SyllableKee

\u{12ac}: 'ኬ'

SyllableKe

\u{12ad}: 'ክ'

SyllableKo

\u{12ae}: 'ኮ'

SyllableKoa

\u{12af}: 'ኯ'

SyllableKwa

\u{12b0}: 'ኰ'

SyllableKwi

\u{12b2}: 'ኲ'

SyllableKwaa

\u{12b3}: 'ኳ'

SyllableKwee

\u{12b4}: 'ኴ'

SyllableKwe

\u{12b5}: 'ኵ'

SyllableKxa

\u{12b8}: 'ኸ'

SyllableKxu

\u{12b9}: 'ኹ'

SyllableKxi

\u{12ba}: 'ኺ'

SyllableKxaa

\u{12bb}: 'ኻ'

SyllableKxee

\u{12bc}: 'ኼ'

SyllableKxe

\u{12bd}: 'ኽ'

SyllableKxo

\u{12be}: 'ኾ'

SyllableKxwa

\u{12c0}: 'ዀ'

SyllableKxwi

\u{12c2}: 'ዂ'

SyllableKxwaa

\u{12c3}: 'ዃ'

SyllableKxwee

\u{12c4}: 'ዄ'

SyllableKxwe

\u{12c5}: 'ዅ'

SyllableWa

\u{12c8}: 'ወ'

SyllableWu

\u{12c9}: 'ዉ'

SyllableWi

\u{12ca}: 'ዊ'

SyllableWaa

\u{12cb}: 'ዋ'

SyllableWee

\u{12cc}: 'ዌ'

SyllableWe

\u{12cd}: 'ው'

SyllableWo

\u{12ce}: 'ዎ'

SyllableWoa

\u{12cf}: 'ዏ'

SyllablePharyngealA

\u{12d0}: 'ዐ'

SyllablePharyngealU

\u{12d1}: 'ዑ'

SyllablePharyngealI

\u{12d2}: 'ዒ'

SyllablePharyngealAa

\u{12d3}: 'ዓ'

SyllablePharyngealEe

\u{12d4}: 'ዔ'

SyllablePharyngealE

\u{12d5}: 'ዕ'

SyllablePharyngealO

\u{12d6}: 'ዖ'

SyllableZa

\u{12d8}: 'ዘ'

SyllableZu

\u{12d9}: 'ዙ'

SyllableZi

\u{12da}: 'ዚ'

SyllableZaa

\u{12db}: 'ዛ'

SyllableZee

\u{12dc}: 'ዜ'

SyllableZe

\u{12dd}: 'ዝ'

SyllableZo

\u{12de}: 'ዞ'

SyllableZwa

\u{12df}: 'ዟ'

SyllableZha

\u{12e0}: 'ዠ'

SyllableZhu

\u{12e1}: 'ዡ'

SyllableZhi

\u{12e2}: 'ዢ'

SyllableZhaa

\u{12e3}: 'ዣ'

SyllableZhee

\u{12e4}: 'ዤ'

SyllableZhe

\u{12e5}: 'ዥ'

SyllableZho

\u{12e6}: 'ዦ'

SyllableZhwa

\u{12e7}: 'ዧ'

SyllableYa

\u{12e8}: 'የ'

SyllableYu

\u{12e9}: 'ዩ'

SyllableYi

\u{12ea}: 'ዪ'

SyllableYaa

\u{12eb}: 'ያ'

SyllableYee

\u{12ec}: 'ዬ'

SyllableYe

\u{12ed}: 'ይ'

SyllableYo

\u{12ee}: 'ዮ'

SyllableYoa

\u{12ef}: 'ዯ'

SyllableDa

\u{12f0}: 'ደ'

SyllableDu

\u{12f1}: 'ዱ'

SyllableDi

\u{12f2}: 'ዲ'

SyllableDaa

\u{12f3}: 'ዳ'

SyllableDee

\u{12f4}: 'ዴ'

SyllableDe

\u{12f5}: 'ድ'

SyllableDo

\u{12f6}: 'ዶ'

SyllableDwa

\u{12f7}: 'ዷ'

SyllableDda

\u{12f8}: 'ዸ'

SyllableDdu

\u{12f9}: 'ዹ'

SyllableDdi

\u{12fa}: 'ዺ'

SyllableDdaa

\u{12fb}: 'ዻ'

SyllableDdee

\u{12fc}: 'ዼ'

SyllableDde

\u{12fd}: 'ዽ'

SyllableDdo

\u{12fe}: 'ዾ'

SyllableDdwa

\u{12ff}: 'ዿ'

SyllableJa

\u{1300}: 'ጀ'

SyllableJu

\u{1301}: 'ጁ'

SyllableJi

\u{1302}: 'ጂ'

SyllableJaa

\u{1303}: 'ጃ'

SyllableJee

\u{1304}: 'ጄ'

SyllableJe

\u{1305}: 'ጅ'

SyllableJo

\u{1306}: 'ጆ'

SyllableJwa

\u{1307}: 'ጇ'

SyllableGa

\u{1308}: 'ገ'

SyllableGu

\u{1309}: 'ጉ'

SyllableGi

\u{130a}: 'ጊ'

SyllableGaa

\u{130b}: 'ጋ'

SyllableGee

\u{130c}: 'ጌ'

SyllableGe

\u{130d}: 'ግ'

SyllableGo

\u{130e}: 'ጎ'

SyllableGoa

\u{130f}: 'ጏ'

SyllableGwa

\u{1310}: 'ጐ'

SyllableGwi

\u{1312}: 'ጒ'

SyllableGwaa

\u{1313}: 'ጓ'

SyllableGwee

\u{1314}: 'ጔ'

SyllableGwe

\u{1315}: 'ጕ'

SyllableGga

\u{1318}: 'ጘ'

SyllableGgu

\u{1319}: 'ጙ'

SyllableGgi

\u{131a}: 'ጚ'

SyllableGgaa

\u{131b}: 'ጛ'

SyllableGgee

\u{131c}: 'ጜ'

SyllableGge

\u{131d}: 'ጝ'

SyllableGgo

\u{131e}: 'ጞ'

SyllableGgwaa

\u{131f}: 'ጟ'

SyllableTha

\u{1320}: 'ጠ'

SyllableThu

\u{1321}: 'ጡ'

SyllableThi

\u{1322}: 'ጢ'

SyllableThaa

\u{1323}: 'ጣ'

SyllableThee

\u{1324}: 'ጤ'

SyllableThe

\u{1325}: 'ጥ'

SyllableTho

\u{1326}: 'ጦ'

SyllableThwa

\u{1327}: 'ጧ'

SyllableCha

\u{1328}: 'ጨ'

SyllableChu

\u{1329}: 'ጩ'

SyllableChi

\u{132a}: 'ጪ'

SyllableChaa

\u{132b}: 'ጫ'

SyllableChee

\u{132c}: 'ጬ'

SyllableChe

\u{132d}: 'ጭ'

SyllableCho

\u{132e}: 'ጮ'

SyllableChwa

\u{132f}: 'ጯ'

SyllablePha

\u{1330}: 'ጰ'

SyllablePhu

\u{1331}: 'ጱ'

SyllablePhi

\u{1332}: 'ጲ'

SyllablePhaa

\u{1333}: 'ጳ'

SyllablePhee

\u{1334}: 'ጴ'

SyllablePhe

\u{1335}: 'ጵ'

SyllablePho

\u{1336}: 'ጶ'

SyllablePhwa

\u{1337}: 'ጷ'

SyllableTsa

\u{1338}: 'ጸ'

SyllableTsu

\u{1339}: 'ጹ'

SyllableTsi

\u{133a}: 'ጺ'

SyllableTsaa

\u{133b}: 'ጻ'

SyllableTsee

\u{133c}: 'ጼ'

SyllableTse

\u{133d}: 'ጽ'

SyllableTso

\u{133e}: 'ጾ'

SyllableTswa

\u{133f}: 'ጿ'

SyllableTza

\u{1340}: 'ፀ'

SyllableTzu

\u{1341}: 'ፁ'

SyllableTzi

\u{1342}: 'ፂ'

SyllableTzaa

\u{1343}: 'ፃ'

SyllableTzee

\u{1344}: 'ፄ'

SyllableTze

\u{1345}: 'ፅ'

SyllableTzo

\u{1346}: 'ፆ'

SyllableTzoa

\u{1347}: 'ፇ'

SyllableFa

\u{1348}: 'ፈ'

SyllableFu

\u{1349}: 'ፉ'

SyllableFi

\u{134a}: 'ፊ'

SyllableFaa

\u{134b}: 'ፋ'

SyllableFee

\u{134c}: 'ፌ'

SyllableFe

\u{134d}: 'ፍ'

SyllableFo

\u{134e}: 'ፎ'

SyllableFwa

\u{134f}: 'ፏ'

SyllablePa

\u{1350}: 'ፐ'

SyllablePu

\u{1351}: 'ፑ'

SyllablePi

\u{1352}: 'ፒ'

SyllablePaa

\u{1353}: 'ፓ'

SyllablePee

\u{1354}: 'ፔ'

SyllablePe

\u{1355}: 'ፕ'

SyllablePo

\u{1356}: 'ፖ'

SyllablePwa

\u{1357}: 'ፗ'

SyllableRya

\u{1358}: 'ፘ'

SyllableMya

\u{1359}: 'ፙ'

SyllableFya

\u{135a}: 'ፚ'

CombiningGeminationAndVowelLengthMark

\u{135d}: '፝'

CombiningVowelLengthMark

\u{135e}: '፞'

CombiningGeminationMark

\u{135f}: '፟'

SectionMark

\u{1360}: '፠'

Wordspace

\u{1361}: '፡'

FullStop

\u{1362}: '።'

Comma

\u{1363}: '፣'

Semicolon

\u{1364}: '፤'

Colon

\u{1365}: '፥'

PrefaceColon

\u{1366}: '፦'

QuestionMark

\u{1367}: '፧'

ParagraphSeparator

\u{1368}: '፨'

DigitOne

\u{1369}: '፩'

DigitTwo

\u{136a}: '፪'

DigitThree

\u{136b}: '፫'

DigitFour

\u{136c}: '፬'

DigitFive

\u{136d}: '፭'

DigitSix

\u{136e}: '፮'

DigitSeven

\u{136f}: '፯'

DigitEight

\u{1370}: '፰'

DigitNine

\u{1371}: '፱'

NumberTen

\u{1372}: '፲'

NumberTwenty

\u{1373}: '፳'

NumberThirty

\u{1374}: '፴'

NumberForty

\u{1375}: '፵'

NumberFifty

\u{1376}: '፶'

NumberSixty

\u{1377}: '፷'

NumberSeventy

\u{1378}: '፸'

NumberEighty

\u{1379}: '፹'

NumberNinety

\u{137a}: '፺'

NumberHundred

\u{137b}: '፻'

NumberTenThousand

\u{137c}: '፼'

Methods

impl Ethiopic

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case