Enum unicode_types::glagolitic::Glagolitic

pub enum Glagolitic {
    CapitalLetterAzu,
    CapitalLetterBuky,
    CapitalLetterVede,
    CapitalLetterGlagoli,
    CapitalLetterDobro,
    CapitalLetterYestu,
    CapitalLetterZhivete,
    CapitalLetterDzelo,
    CapitalLetterZemlja,
    CapitalLetterIzhe,
    CapitalLetterInitialIzhe,
    CapitalLetterI,
    CapitalLetterDjervi,
    CapitalLetterKako,
    CapitalLetterLjudije,
    CapitalLetterMyslite,
    CapitalLetterNashi,
    CapitalLetterOnu,
    CapitalLetterPokoji,
    CapitalLetterRitsi,
    CapitalLetterSlovo,
    CapitalLetterTvrido,
    CapitalLetterUku,
    CapitalLetterFritu,
    CapitalLetterHeru,
    CapitalLetterOtu,
    CapitalLetterPe,
    CapitalLetterShta,
    CapitalLetterTsi,
    CapitalLetterChrivi,
    CapitalLetterSha,
    CapitalLetterYeru,
    CapitalLetterYeri,
    CapitalLetterYati,
    CapitalLetterSpideryHa,
    CapitalLetterYu,
    CapitalLetterSmallYus,
    CapitalLetterSmallYusWithTail,
    CapitalLetterYo,
    CapitalLetterIotatedSmallYus,
    CapitalLetterBigYus,
    CapitalLetterIotatedBigYus,
    CapitalLetterFita,
    CapitalLetterIzhitsa,
    CapitalLetterShtapic,
    CapitalLetterTrokutastiA,
    CapitalLetterLatinateMyslite,
    SmallLetterAzu,
    SmallLetterBuky,
    SmallLetterVede,
    SmallLetterGlagoli,
    SmallLetterDobro,
    SmallLetterYestu,
    SmallLetterZhivete,
    SmallLetterDzelo,
    SmallLetterZemlja,
    SmallLetterIzhe,
    SmallLetterInitialIzhe,
    SmallLetterI,
    SmallLetterDjervi,
    SmallLetterKako,
    SmallLetterLjudije,
    SmallLetterMyslite,
    SmallLetterNashi,
    SmallLetterOnu,
    SmallLetterPokoji,
    SmallLetterRitsi,
    SmallLetterSlovo,
    SmallLetterTvrido,
    SmallLetterUku,
    SmallLetterFritu,
    SmallLetterHeru,
    SmallLetterOtu,
    SmallLetterPe,
    SmallLetterShta,
    SmallLetterTsi,
    SmallLetterChrivi,
    SmallLetterSha,
    SmallLetterYeru,
    SmallLetterYeri,
    SmallLetterYati,
    SmallLetterSpideryHa,
    SmallLetterYu,
    SmallLetterSmallYus,
    SmallLetterSmallYusWithTail,
    SmallLetterYo,
    SmallLetterIotatedSmallYus,
    SmallLetterBigYus,
    SmallLetterIotatedBigYus,
    SmallLetterFita,
    SmallLetterIzhitsa,
    SmallLetterShtapic,
    SmallLetterTrokutastiA,
    SmallLetterLatinateMyslite,
}

An enum to represent all characters in the Glagolitic block.

Variants
CapitalLetterAzu

\u{2c00}: 'Ⰰ'

CapitalLetterBuky

\u{2c01}: 'Ⰱ'

CapitalLetterVede

\u{2c02}: 'Ⰲ'

CapitalLetterGlagoli

\u{2c03}: 'Ⰳ'

CapitalLetterDobro

\u{2c04}: 'Ⰴ'

CapitalLetterYestu

\u{2c05}: 'Ⰵ'

CapitalLetterZhivete

\u{2c06}: 'Ⰶ'

CapitalLetterDzelo

\u{2c07}: 'Ⰷ'

CapitalLetterZemlja

\u{2c08}: 'Ⰸ'

CapitalLetterIzhe

\u{2c09}: 'Ⰹ'

CapitalLetterInitialIzhe

\u{2c0a}: 'Ⰺ'

CapitalLetterI

\u{2c0b}: 'Ⰻ'

CapitalLetterDjervi

\u{2c0c}: 'Ⰼ'

CapitalLetterKako

\u{2c0d}: 'Ⰽ'

CapitalLetterLjudije

\u{2c0e}: 'Ⰾ'

CapitalLetterMyslite

\u{2c0f}: 'Ⰿ'

CapitalLetterNashi

\u{2c10}: 'Ⱀ'

CapitalLetterOnu

\u{2c11}: 'Ⱁ'

CapitalLetterPokoji

\u{2c12}: 'Ⱂ'

CapitalLetterRitsi

\u{2c13}: 'Ⱃ'

CapitalLetterSlovo

\u{2c14}: 'Ⱄ'

CapitalLetterTvrido

\u{2c15}: 'Ⱅ'

CapitalLetterUku

\u{2c16}: 'Ⱆ'

CapitalLetterFritu

\u{2c17}: 'Ⱇ'

CapitalLetterHeru

\u{2c18}: 'Ⱈ'

CapitalLetterOtu

\u{2c19}: 'Ⱉ'

CapitalLetterPe

\u{2c1a}: 'Ⱊ'

CapitalLetterShta

\u{2c1b}: 'Ⱋ'

CapitalLetterTsi

\u{2c1c}: 'Ⱌ'

CapitalLetterChrivi

\u{2c1d}: 'Ⱍ'

CapitalLetterSha

\u{2c1e}: 'Ⱎ'

CapitalLetterYeru

\u{2c1f}: 'Ⱏ'

CapitalLetterYeri

\u{2c20}: 'Ⱐ'

CapitalLetterYati

\u{2c21}: 'Ⱑ'

CapitalLetterSpideryHa

\u{2c22}: 'Ⱒ'

CapitalLetterYu

\u{2c23}: 'Ⱓ'

CapitalLetterSmallYus

\u{2c24}: 'Ⱔ'

CapitalLetterSmallYusWithTail

\u{2c25}: 'Ⱕ'

CapitalLetterYo

\u{2c26}: 'Ⱖ'

CapitalLetterIotatedSmallYus

\u{2c27}: 'Ⱗ'

CapitalLetterBigYus

\u{2c28}: 'Ⱘ'

CapitalLetterIotatedBigYus

\u{2c29}: 'Ⱙ'

CapitalLetterFita

\u{2c2a}: 'Ⱚ'

CapitalLetterIzhitsa

\u{2c2b}: 'Ⱛ'

CapitalLetterShtapic

\u{2c2c}: 'Ⱜ'

CapitalLetterTrokutastiA

\u{2c2d}: 'Ⱝ'

CapitalLetterLatinateMyslite

\u{2c2e}: 'Ⱞ'

SmallLetterAzu

\u{2c30}: 'ⰰ'

SmallLetterBuky

\u{2c31}: 'ⰱ'

SmallLetterVede

\u{2c32}: 'ⰲ'

SmallLetterGlagoli

\u{2c33}: 'ⰳ'

SmallLetterDobro

\u{2c34}: 'ⰴ'

SmallLetterYestu

\u{2c35}: 'ⰵ'

SmallLetterZhivete

\u{2c36}: 'ⰶ'

SmallLetterDzelo

\u{2c37}: 'ⰷ'

SmallLetterZemlja

\u{2c38}: 'ⰸ'

SmallLetterIzhe

\u{2c39}: 'ⰹ'

SmallLetterInitialIzhe

\u{2c3a}: 'ⰺ'

SmallLetterI

\u{2c3b}: 'ⰻ'

SmallLetterDjervi

\u{2c3c}: 'ⰼ'

SmallLetterKako

\u{2c3d}: 'ⰽ'

SmallLetterLjudije

\u{2c3e}: 'ⰾ'

SmallLetterMyslite

\u{2c3f}: 'ⰿ'

SmallLetterNashi

\u{2c40}: 'ⱀ'

SmallLetterOnu

\u{2c41}: 'ⱁ'

SmallLetterPokoji

\u{2c42}: 'ⱂ'

SmallLetterRitsi

\u{2c43}: 'ⱃ'

SmallLetterSlovo

\u{2c44}: 'ⱄ'

SmallLetterTvrido

\u{2c45}: 'ⱅ'

SmallLetterUku

\u{2c46}: 'ⱆ'

SmallLetterFritu

\u{2c47}: 'ⱇ'

SmallLetterHeru

\u{2c48}: 'ⱈ'

SmallLetterOtu

\u{2c49}: 'ⱉ'

SmallLetterPe

\u{2c4a}: 'ⱊ'

SmallLetterShta

\u{2c4b}: 'ⱋ'

SmallLetterTsi

\u{2c4c}: 'ⱌ'

SmallLetterChrivi

\u{2c4d}: 'ⱍ'

SmallLetterSha

\u{2c4e}: 'ⱎ'

SmallLetterYeru

\u{2c4f}: 'ⱏ'

SmallLetterYeri

\u{2c50}: 'ⱐ'

SmallLetterYati

\u{2c51}: 'ⱑ'

SmallLetterSpideryHa

\u{2c52}: 'ⱒ'

SmallLetterYu

\u{2c53}: 'ⱓ'

SmallLetterSmallYus

\u{2c54}: 'ⱔ'

SmallLetterSmallYusWithTail

\u{2c55}: 'ⱕ'

SmallLetterYo

\u{2c56}: 'ⱖ'

SmallLetterIotatedSmallYus

\u{2c57}: 'ⱗ'

SmallLetterBigYus

\u{2c58}: 'ⱘ'

SmallLetterIotatedBigYus

\u{2c59}: 'ⱙ'

SmallLetterFita

\u{2c5a}: 'ⱚ'

SmallLetterIzhitsa

\u{2c5b}: 'ⱛ'

SmallLetterShtapic

\u{2c5c}: 'ⱜ'

SmallLetterTrokutastiA

\u{2c5d}: 'ⱝ'

SmallLetterLatinateMyslite

\u{2c5e}: 'ⱞ'

Methods

impl Glagolitic

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case