Enum unicode_types::greek_and_coptic::GreekandCoptic

pub enum GreekandCoptic {
    GreekCapitalLetterHeta,
    GreekSmallLetterHeta,
    GreekCapitalLetterArchaicSampi,
    GreekSmallLetterArchaicSampi,
    GreekNumeralSign,
    GreekLowerNumeralSign,
    GreekCapitalLetterPamphylianDigamma,
    GreekSmallLetterPamphylianDigamma,
    GreekYpogegrammeni,
    GreekSmallReversedLunateSigmaSymbol,
    GreekSmallDottedLunateSigmaSymbol,
    GreekSmallReversedDottedLunateSigmaSymbol,
    GreekQuestionMark,
    GreekCapitalLetterYot,
    GreekTonos,
    GreekDialytikaTonos,
    GreekCapitalLetterAlphaWithTonos,
    GreekAnoTeleia,
    GreekCapitalLetterEpsilonWithTonos,
    GreekCapitalLetterEtaWithTonos,
    GreekCapitalLetterIotaWithTonos,
    GreekCapitalLetterOmicronWithTonos,
    GreekCapitalLetterUpsilonWithTonos,
    GreekCapitalLetterOmegaWithTonos,
    GreekSmallLetterIotaWithDialytikaAndTonos,
    GreekCapitalLetterAlpha,
    GreekCapitalLetterBeta,
    GreekCapitalLetterGamma,
    GreekCapitalLetterDelta,
    GreekCapitalLetterEpsilon,
    GreekCapitalLetterZeta,
    GreekCapitalLetterEta,
    GreekCapitalLetterTheta,
    GreekCapitalLetterIota,
    GreekCapitalLetterKappa,
    GreekCapitalLetterLamda,
    GreekCapitalLetterMu,
    GreekCapitalLetterNu,
    GreekCapitalLetterXi,
    GreekCapitalLetterOmicron,
    GreekCapitalLetterPi,
    GreekCapitalLetterRho,
    GreekCapitalLetterSigma,
    GreekCapitalLetterTau,
    GreekCapitalLetterUpsilon,
    GreekCapitalLetterPhi,
    GreekCapitalLetterChi,
    GreekCapitalLetterPsi,
    GreekCapitalLetterOmega,
    GreekCapitalLetterIotaWithDialytika,
    GreekCapitalLetterUpsilonWithDialytika,
    GreekSmallLetterAlphaWithTonos,
    GreekSmallLetterEpsilonWithTonos,
    GreekSmallLetterEtaWithTonos,
    GreekSmallLetterIotaWithTonos,
    GreekSmallLetterUpsilonWithDialytikaAndTonos,
    GreekSmallLetterAlpha,
    GreekSmallLetterBeta,
    GreekSmallLetterGamma,
    GreekSmallLetterDelta,
    GreekSmallLetterEpsilon,
    GreekSmallLetterZeta,
    GreekSmallLetterEta,
    GreekSmallLetterTheta,
    GreekSmallLetterIota,
    GreekSmallLetterKappa,
    GreekSmallLetterLamda,
    GreekSmallLetterMu,
    GreekSmallLetterNu,
    GreekSmallLetterXi,
    GreekSmallLetterOmicron,
    GreekSmallLetterPi,
    GreekSmallLetterRho,
    GreekSmallLetterFinalSigma,
    GreekSmallLetterSigma,
    GreekSmallLetterTau,
    GreekSmallLetterUpsilon,
    GreekSmallLetterPhi,
    GreekSmallLetterChi,
    GreekSmallLetterPsi,
    GreekSmallLetterOmega,
    GreekSmallLetterIotaWithDialytika,
    GreekSmallLetterUpsilonWithDialytika,
    GreekSmallLetterOmicronWithTonos,
    GreekSmallLetterUpsilonWithTonos,
    GreekSmallLetterOmegaWithTonos,
    GreekCapitalKaiSymbol,
    GreekBetaSymbol,
    GreekThetaSymbol,
    GreekUpsilonWithHookSymbol,
    GreekUpsilonWithAcuteAndHookSymbol,
    GreekUpsilonWithDiaeresisAndHookSymbol,
    GreekPhiSymbol,
    GreekPiSymbol,
    GreekKaiSymbol,
    GreekLetterArchaicKoppa,
    GreekSmallLetterArchaicKoppa,
    GreekLetterStigma,
    GreekSmallLetterStigma,
    GreekLetterDigamma,
    GreekSmallLetterDigamma,
    GreekLetterKoppa,
    GreekSmallLetterKoppa,
    GreekLetterSampi,
    GreekSmallLetterSampi,
    CopticCapitalLetterShei,
    CopticSmallLetterShei,
    CopticCapitalLetterFei,
    CopticSmallLetterFei,
    CopticCapitalLetterKhei,
    CopticSmallLetterKhei,
    CopticCapitalLetterHori,
    CopticSmallLetterHori,
    CopticCapitalLetterGangia,
    CopticSmallLetterGangia,
    CopticCapitalLetterShima,
    CopticSmallLetterShima,
    CopticCapitalLetterDei,
    CopticSmallLetterDei,
    GreekKappaSymbol,
    GreekRhoSymbol,
    GreekLunateSigmaSymbol,
    GreekLetterYot,
    GreekCapitalThetaSymbol,
    GreekLunateEpsilonSymbol,
    GreekReversedLunateEpsilonSymbol,
    GreekCapitalLetterSho,
    GreekSmallLetterSho,
    GreekCapitalLunateSigmaSymbol,
    GreekCapitalLetterSan,
    GreekSmallLetterSan,
    GreekRhoWithStrokeSymbol,
    GreekCapitalReversedLunateSigmaSymbol,
    GreekCapitalDottedLunateSigmaSymbol,
}

An enum to represent all characters in the GreekandCoptic block.

Variants
GreekCapitalLetterHeta

\u{370}: 'Ͱ'

GreekSmallLetterHeta

\u{371}: 'ͱ'

GreekCapitalLetterArchaicSampi

\u{372}: 'Ͳ'

GreekSmallLetterArchaicSampi

\u{373}: 'ͳ'

GreekNumeralSign

\u{374}: 'ʹ'

GreekLowerNumeralSign

\u{375}: '͵'

GreekCapitalLetterPamphylianDigamma

\u{376}: 'Ͷ'

GreekSmallLetterPamphylianDigamma

\u{377}: 'ͷ'

GreekYpogegrammeni

\u{37a}: 'ͺ'

GreekSmallReversedLunateSigmaSymbol

\u{37b}: 'ͻ'

GreekSmallDottedLunateSigmaSymbol

\u{37c}: 'ͼ'

GreekSmallReversedDottedLunateSigmaSymbol

\u{37d}: 'ͽ'

GreekQuestionMark

\u{37e}: ';'

GreekCapitalLetterYot

\u{37f}: 'Ϳ'

GreekTonos

\u{384}: '΄'

GreekDialytikaTonos

\u{385}: '΅'

GreekCapitalLetterAlphaWithTonos

\u{386}: 'Ά'

GreekAnoTeleia

\u{387}: '·'

GreekCapitalLetterEpsilonWithTonos

\u{388}: 'Έ'

GreekCapitalLetterEtaWithTonos

\u{389}: 'Ή'

GreekCapitalLetterIotaWithTonos

\u{38a}: 'Ί'

GreekCapitalLetterOmicronWithTonos

\u{38c}: 'Ό'

GreekCapitalLetterUpsilonWithTonos

\u{38e}: 'Ύ'

GreekCapitalLetterOmegaWithTonos

\u{38f}: 'Ώ'

GreekSmallLetterIotaWithDialytikaAndTonos

\u{390}: 'ΐ'

GreekCapitalLetterAlpha

\u{391}: 'Α'

GreekCapitalLetterBeta

\u{392}: 'Β'

GreekCapitalLetterGamma

\u{393}: 'Γ'

GreekCapitalLetterDelta

\u{394}: 'Δ'

GreekCapitalLetterEpsilon

\u{395}: 'Ε'

GreekCapitalLetterZeta

\u{396}: 'Ζ'

GreekCapitalLetterEta

\u{397}: 'Η'

GreekCapitalLetterTheta

\u{398}: 'Θ'

GreekCapitalLetterIota

\u{399}: 'Ι'

GreekCapitalLetterKappa

\u{39a}: 'Κ'

GreekCapitalLetterLamda

\u{39b}: 'Λ'

GreekCapitalLetterMu

\u{39c}: 'Μ'

GreekCapitalLetterNu

\u{39d}: 'Ν'

GreekCapitalLetterXi

\u{39e}: 'Ξ'

GreekCapitalLetterOmicron

\u{39f}: 'Ο'

GreekCapitalLetterPi

\u{3a0}: 'Π'

GreekCapitalLetterRho

\u{3a1}: 'Ρ'

GreekCapitalLetterSigma

\u{3a3}: 'Σ'

GreekCapitalLetterTau

\u{3a4}: 'Τ'

GreekCapitalLetterUpsilon

\u{3a5}: 'Υ'

GreekCapitalLetterPhi

\u{3a6}: 'Φ'

GreekCapitalLetterChi

\u{3a7}: 'Χ'

GreekCapitalLetterPsi

\u{3a8}: 'Ψ'

GreekCapitalLetterOmega

\u{3a9}: 'Ω'

GreekCapitalLetterIotaWithDialytika

\u{3aa}: 'Ϊ'

GreekCapitalLetterUpsilonWithDialytika

\u{3ab}: 'Ϋ'

GreekSmallLetterAlphaWithTonos

\u{3ac}: 'ά'

GreekSmallLetterEpsilonWithTonos

\u{3ad}: 'έ'

GreekSmallLetterEtaWithTonos

\u{3ae}: 'ή'

GreekSmallLetterIotaWithTonos

\u{3af}: 'ί'

GreekSmallLetterUpsilonWithDialytikaAndTonos

\u{3b0}: 'ΰ'

GreekSmallLetterAlpha

\u{3b1}: 'α'

GreekSmallLetterBeta

\u{3b2}: 'β'

GreekSmallLetterGamma

\u{3b3}: 'γ'

GreekSmallLetterDelta

\u{3b4}: 'δ'

GreekSmallLetterEpsilon

\u{3b5}: 'ε'

GreekSmallLetterZeta

\u{3b6}: 'ζ'

GreekSmallLetterEta

\u{3b7}: 'η'

GreekSmallLetterTheta

\u{3b8}: 'θ'

GreekSmallLetterIota

\u{3b9}: 'ι'

GreekSmallLetterKappa

\u{3ba}: 'κ'

GreekSmallLetterLamda

\u{3bb}: 'λ'

GreekSmallLetterMu

\u{3bc}: 'μ'

GreekSmallLetterNu

\u{3bd}: 'ν'

GreekSmallLetterXi

\u{3be}: 'ξ'

GreekSmallLetterOmicron

\u{3bf}: 'ο'

GreekSmallLetterPi

\u{3c0}: 'π'

GreekSmallLetterRho

\u{3c1}: 'ρ'

GreekSmallLetterFinalSigma

\u{3c2}: 'ς'

GreekSmallLetterSigma

\u{3c3}: 'σ'

GreekSmallLetterTau

\u{3c4}: 'τ'

GreekSmallLetterUpsilon

\u{3c5}: 'υ'

GreekSmallLetterPhi

\u{3c6}: 'φ'

GreekSmallLetterChi

\u{3c7}: 'χ'

GreekSmallLetterPsi

\u{3c8}: 'ψ'

GreekSmallLetterOmega

\u{3c9}: 'ω'

GreekSmallLetterIotaWithDialytika

\u{3ca}: 'ϊ'

GreekSmallLetterUpsilonWithDialytika

\u{3cb}: 'ϋ'

GreekSmallLetterOmicronWithTonos

\u{3cc}: 'ό'

GreekSmallLetterUpsilonWithTonos

\u{3cd}: 'ύ'

GreekSmallLetterOmegaWithTonos

\u{3ce}: 'ώ'

GreekCapitalKaiSymbol

\u{3cf}: 'Ϗ'

GreekBetaSymbol

\u{3d0}: 'ϐ'

GreekThetaSymbol

\u{3d1}: 'ϑ'

GreekUpsilonWithHookSymbol

\u{3d2}: 'ϒ'

GreekUpsilonWithAcuteAndHookSymbol

\u{3d3}: 'ϓ'

GreekUpsilonWithDiaeresisAndHookSymbol

\u{3d4}: 'ϔ'

GreekPhiSymbol

\u{3d5}: 'ϕ'

GreekPiSymbol

\u{3d6}: 'ϖ'

GreekKaiSymbol

\u{3d7}: 'ϗ'

GreekLetterArchaicKoppa

\u{3d8}: 'Ϙ'

GreekSmallLetterArchaicKoppa

\u{3d9}: 'ϙ'

GreekLetterStigma

\u{3da}: 'Ϛ'

GreekSmallLetterStigma

\u{3db}: 'ϛ'

GreekLetterDigamma

\u{3dc}: 'Ϝ'

GreekSmallLetterDigamma

\u{3dd}: 'ϝ'

GreekLetterKoppa

\u{3de}: 'Ϟ'

GreekSmallLetterKoppa

\u{3df}: 'ϟ'

GreekLetterSampi

\u{3e0}: 'Ϡ'

GreekSmallLetterSampi

\u{3e1}: 'ϡ'

CopticCapitalLetterShei

\u{3e2}: 'Ϣ'

CopticSmallLetterShei

\u{3e3}: 'ϣ'

CopticCapitalLetterFei

\u{3e4}: 'Ϥ'

CopticSmallLetterFei

\u{3e5}: 'ϥ'

CopticCapitalLetterKhei

\u{3e6}: 'Ϧ'

CopticSmallLetterKhei

\u{3e7}: 'ϧ'

CopticCapitalLetterHori

\u{3e8}: 'Ϩ'

CopticSmallLetterHori

\u{3e9}: 'ϩ'

CopticCapitalLetterGangia

\u{3ea}: 'Ϫ'

CopticSmallLetterGangia

\u{3eb}: 'ϫ'

CopticCapitalLetterShima

\u{3ec}: 'Ϭ'

CopticSmallLetterShima

\u{3ed}: 'ϭ'

CopticCapitalLetterDei

\u{3ee}: 'Ϯ'

CopticSmallLetterDei

\u{3ef}: 'ϯ'

GreekKappaSymbol

\u{3f0}: 'ϰ'

GreekRhoSymbol

\u{3f1}: 'ϱ'

GreekLunateSigmaSymbol

\u{3f2}: 'ϲ'

GreekLetterYot

\u{3f3}: 'ϳ'

GreekCapitalThetaSymbol

\u{3f4}: 'ϴ'

GreekLunateEpsilonSymbol

\u{3f5}: 'ϵ'

GreekReversedLunateEpsilonSymbol

\u{3f6}: '϶'

GreekCapitalLetterSho

\u{3f7}: 'Ϸ'

GreekSmallLetterSho

\u{3f8}: 'ϸ'

GreekCapitalLunateSigmaSymbol

\u{3f9}: 'Ϲ'

GreekCapitalLetterSan

\u{3fa}: 'Ϻ'

GreekSmallLetterSan

\u{3fb}: 'ϻ'

GreekRhoWithStrokeSymbol

\u{3fc}: 'ϼ'

GreekCapitalReversedLunateSigmaSymbol

\u{3fd}: 'Ͻ'

GreekCapitalDottedLunateSigmaSymbol

\u{3fe}: 'Ͼ'

Methods

impl GreekandCoptic

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case