Enum unicode_types::greek_extended::GreekExtended

pub enum GreekExtended {
    GreekSmallLetterAlphaWithPsili,
    GreekSmallLetterAlphaWithDasia,
    GreekSmallLetterAlphaWithPsiliAndVaria,
    GreekSmallLetterAlphaWithDasiaAndVaria,
    GreekSmallLetterAlphaWithPsiliAndOxia,
    GreekSmallLetterAlphaWithDasiaAndOxia,
    GreekSmallLetterAlphaWithPsiliAndPerispomeni,
    GreekSmallLetterAlphaWithDasiaAndPerispomeni,
    GreekCapitalLetterAlphaWithPsili,
    GreekCapitalLetterAlphaWithDasia,
    GreekCapitalLetterAlphaWithPsiliAndVaria,
    GreekCapitalLetterAlphaWithDasiaAndVaria,
    GreekCapitalLetterAlphaWithPsiliAndOxia,
    GreekCapitalLetterAlphaWithDasiaAndOxia,
    GreekCapitalLetterAlphaWithPsiliAndPerispomeni,
    GreekCapitalLetterAlphaWithDasiaAndPerispomeni,
    GreekSmallLetterEpsilonWithPsili,
    GreekSmallLetterEpsilonWithDasia,
    GreekSmallLetterEpsilonWithPsiliAndVaria,
    GreekSmallLetterEpsilonWithDasiaAndVaria,
    GreekSmallLetterEpsilonWithPsiliAndOxia,
    GreekSmallLetterEpsilonWithDasiaAndOxia,
    GreekCapitalLetterEpsilonWithPsili,
    GreekCapitalLetterEpsilonWithDasia,
    GreekCapitalLetterEpsilonWithPsiliAndVaria,
    GreekCapitalLetterEpsilonWithDasiaAndVaria,
    GreekCapitalLetterEpsilonWithPsiliAndOxia,
    GreekCapitalLetterEpsilonWithDasiaAndOxia,
    GreekSmallLetterEtaWithPsili,
    GreekSmallLetterEtaWithDasia,
    GreekSmallLetterEtaWithPsiliAndVaria,
    GreekSmallLetterEtaWithDasiaAndVaria,
    GreekSmallLetterEtaWithPsiliAndOxia,
    GreekSmallLetterEtaWithDasiaAndOxia,
    GreekSmallLetterEtaWithPsiliAndPerispomeni,
    GreekSmallLetterEtaWithDasiaAndPerispomeni,
    GreekCapitalLetterEtaWithPsili,
    GreekCapitalLetterEtaWithDasia,
    GreekCapitalLetterEtaWithPsiliAndVaria,
    GreekCapitalLetterEtaWithDasiaAndVaria,
    GreekCapitalLetterEtaWithPsiliAndOxia,
    GreekCapitalLetterEtaWithDasiaAndOxia,
    GreekCapitalLetterEtaWithPsiliAndPerispomeni,
    GreekCapitalLetterEtaWithDasiaAndPerispomeni,
    GreekSmallLetterIotaWithPsili,
    GreekSmallLetterIotaWithDasia,
    GreekSmallLetterIotaWithPsiliAndVaria,
    GreekSmallLetterIotaWithDasiaAndVaria,
    GreekSmallLetterIotaWithPsiliAndOxia,
    GreekSmallLetterIotaWithDasiaAndOxia,
    GreekSmallLetterIotaWithPsiliAndPerispomeni,
    GreekSmallLetterIotaWithDasiaAndPerispomeni,
    GreekCapitalLetterIotaWithPsili,
    GreekCapitalLetterIotaWithDasia,
    GreekCapitalLetterIotaWithPsiliAndVaria,
    GreekCapitalLetterIotaWithDasiaAndVaria,
    GreekCapitalLetterIotaWithPsiliAndOxia,
    GreekCapitalLetterIotaWithDasiaAndOxia,
    GreekCapitalLetterIotaWithPsiliAndPerispomeni,
    GreekCapitalLetterIotaWithDasiaAndPerispomeni,
    GreekSmallLetterOmicronWithPsili,
    GreekSmallLetterOmicronWithDasia,
    GreekSmallLetterOmicronWithPsiliAndVaria,
    GreekSmallLetterOmicronWithDasiaAndVaria,
    GreekSmallLetterOmicronWithPsiliAndOxia,
    GreekSmallLetterOmicronWithDasiaAndOxia,
    GreekCapitalLetterOmicronWithPsili,
    GreekCapitalLetterOmicronWithDasia,
    GreekCapitalLetterOmicronWithPsiliAndVaria,
    GreekCapitalLetterOmicronWithDasiaAndVaria,
    GreekCapitalLetterOmicronWithPsiliAndOxia,
    GreekCapitalLetterOmicronWithDasiaAndOxia,
    GreekSmallLetterUpsilonWithPsili,
    GreekSmallLetterUpsilonWithDasia,
    GreekSmallLetterUpsilonWithPsiliAndVaria,
    GreekSmallLetterUpsilonWithDasiaAndVaria,
    GreekSmallLetterUpsilonWithPsiliAndOxia,
    GreekSmallLetterUpsilonWithDasiaAndOxia,
    GreekSmallLetterUpsilonWithPsiliAndPerispomeni,
    GreekSmallLetterUpsilonWithDasiaAndPerispomeni,
    GreekCapitalLetterUpsilonWithDasia,
    GreekCapitalLetterUpsilonWithDasiaAndVaria,
    GreekCapitalLetterUpsilonWithDasiaAndOxia,
    GreekCapitalLetterUpsilonWithDasiaAndPerispomeni,
    GreekSmallLetterOmegaWithPsili,
    GreekSmallLetterOmegaWithDasia,
    GreekSmallLetterOmegaWithPsiliAndVaria,
    GreekSmallLetterOmegaWithDasiaAndVaria,
    GreekSmallLetterOmegaWithPsiliAndOxia,
    GreekSmallLetterOmegaWithDasiaAndOxia,
    GreekSmallLetterOmegaWithPsiliAndPerispomeni,
    GreekSmallLetterOmegaWithDasiaAndPerispomeni,
    GreekCapitalLetterOmegaWithPsili,
    GreekCapitalLetterOmegaWithDasia,
    GreekCapitalLetterOmegaWithPsiliAndVaria,
    GreekCapitalLetterOmegaWithDasiaAndVaria,
    GreekCapitalLetterOmegaWithPsiliAndOxia,
    GreekCapitalLetterOmegaWithDasiaAndOxia,
    GreekCapitalLetterOmegaWithPsiliAndPerispomeni,
    GreekCapitalLetterOmegaWithDasiaAndPerispomeni,
    GreekSmallLetterAlphaWithVaria,
    GreekSmallLetterAlphaWithOxia,
    GreekSmallLetterEpsilonWithVaria,
    GreekSmallLetterEpsilonWithOxia,
    GreekSmallLetterEtaWithVaria,
    GreekSmallLetterEtaWithOxia,
    GreekSmallLetterIotaWithVaria,
    GreekSmallLetterIotaWithOxia,
    GreekSmallLetterOmicronWithVaria,
    GreekSmallLetterOmicronWithOxia,
    GreekSmallLetterUpsilonWithVaria,
    GreekSmallLetterUpsilonWithOxia,
    GreekSmallLetterOmegaWithVaria,
    GreekSmallLetterOmegaWithOxia,
    GreekSmallLetterAlphaWithPsiliAndYpogegrammeni,
    GreekSmallLetterAlphaWithDasiaAndYpogegrammeni,
    GreekSmallLetterAlphaWithPsiliAndVariaAndYpogegrammeni,
    GreekSmallLetterAlphaWithDasiaAndVariaAndYpogegrammeni,
    GreekSmallLetterAlphaWithPsiliAndOxiaAndYpogegrammeni,
    GreekSmallLetterAlphaWithDasiaAndOxiaAndYpogegrammeni,
    GreekSmallLetterAlphaWithPsiliAndPerispomeniAndYpogegrammeni,
    GreekSmallLetterAlphaWithDasiaAndPerispomeniAndYpogegrammeni,
    GreekCapitalLetterAlphaWithPsiliAndProsgegrammeni,
    GreekCapitalLetterAlphaWithDasiaAndProsgegrammeni,
    GreekCapitalLetterAlphaWithPsiliAndVariaAndProsgegrammeni,
    GreekCapitalLetterAlphaWithDasiaAndVariaAndProsgegrammeni,
    GreekCapitalLetterAlphaWithPsiliAndOxiaAndProsgegrammeni,
    GreekCapitalLetterAlphaWithDasiaAndOxiaAndProsgegrammeni,
    GreekCapitalLetterAlphaWithPsiliAndPerispomeniAndProsgegrammeni,
    GreekCapitalLetterAlphaWithDasiaAndPerispomeniAndProsgegrammeni,
    GreekSmallLetterEtaWithPsiliAndYpogegrammeni,
    GreekSmallLetterEtaWithDasiaAndYpogegrammeni,
    GreekSmallLetterEtaWithPsiliAndVariaAndYpogegrammeni,
    GreekSmallLetterEtaWithDasiaAndVariaAndYpogegrammeni,
    GreekSmallLetterEtaWithPsiliAndOxiaAndYpogegrammeni,
    GreekSmallLetterEtaWithDasiaAndOxiaAndYpogegrammeni,
    GreekSmallLetterEtaWithPsiliAndPerispomeniAndYpogegrammeni,
    GreekSmallLetterEtaWithDasiaAndPerispomeniAndYpogegrammeni,
    GreekCapitalLetterEtaWithPsiliAndProsgegrammeni,
    GreekCapitalLetterEtaWithDasiaAndProsgegrammeni,
    GreekCapitalLetterEtaWithPsiliAndVariaAndProsgegrammeni,
    GreekCapitalLetterEtaWithDasiaAndVariaAndProsgegrammeni,
    GreekCapitalLetterEtaWithPsiliAndOxiaAndProsgegrammeni,
    GreekCapitalLetterEtaWithDasiaAndOxiaAndProsgegrammeni,
    GreekCapitalLetterEtaWithPsiliAndPerispomeniAndProsgegrammeni,
    GreekCapitalLetterEtaWithDasiaAndPerispomeniAndProsgegrammeni,
    GreekSmallLetterOmegaWithPsiliAndYpogegrammeni,
    GreekSmallLetterOmegaWithDasiaAndYpogegrammeni,
    GreekSmallLetterOmegaWithPsiliAndVariaAndYpogegrammeni,
    GreekSmallLetterOmegaWithDasiaAndVariaAndYpogegrammeni,
    GreekSmallLetterOmegaWithPsiliAndOxiaAndYpogegrammeni,
    GreekSmallLetterOmegaWithDasiaAndOxiaAndYpogegrammeni,
    GreekSmallLetterOmegaWithPsiliAndPerispomeniAndYpogegrammeni,
    GreekSmallLetterOmegaWithDasiaAndPerispomeniAndYpogegrammeni,
    GreekCapitalLetterOmegaWithPsiliAndProsgegrammeni,
    GreekCapitalLetterOmegaWithDasiaAndProsgegrammeni,
    GreekCapitalLetterOmegaWithPsiliAndVariaAndProsgegrammeni,
    GreekCapitalLetterOmegaWithDasiaAndVariaAndProsgegrammeni,
    GreekCapitalLetterOmegaWithPsiliAndOxiaAndProsgegrammeni,
    GreekCapitalLetterOmegaWithDasiaAndOxiaAndProsgegrammeni,
    GreekCapitalLetterOmegaWithPsiliAndPerispomeniAndProsgegrammeni,
    GreekCapitalLetterOmegaWithDasiaAndPerispomeniAndProsgegrammeni,
    GreekSmallLetterAlphaWithVrachy,
    GreekSmallLetterAlphaWithMacron,
    GreekSmallLetterAlphaWithVariaAndYpogegrammeni,
    GreekSmallLetterAlphaWithYpogegrammeni,
    GreekSmallLetterAlphaWithOxiaAndYpogegrammeni,
    GreekSmallLetterAlphaWithPerispomeni,
    GreekSmallLetterAlphaWithPerispomeniAndYpogegrammeni,
    GreekCapitalLetterAlphaWithVrachy,
    GreekCapitalLetterAlphaWithMacron,
    GreekCapitalLetterAlphaWithVaria,
    GreekCapitalLetterAlphaWithOxia,
    GreekCapitalLetterAlphaWithProsgegrammeni,
    GreekKoronis,
    GreekProsgegrammeni,
    GreekPsili,
    GreekPerispomeni,
    GreekDialytikaAndPerispomeni,
    GreekSmallLetterEtaWithVariaAndYpogegrammeni,
    GreekSmallLetterEtaWithYpogegrammeni,
    GreekSmallLetterEtaWithOxiaAndYpogegrammeni,
    GreekSmallLetterEtaWithPerispomeni,
    GreekSmallLetterEtaWithPerispomeniAndYpogegrammeni,
    GreekCapitalLetterEpsilonWithVaria,
    GreekCapitalLetterEpsilonWithOxia,
    GreekCapitalLetterEtaWithVaria,
    GreekCapitalLetterEtaWithOxia,
    GreekCapitalLetterEtaWithProsgegrammeni,
    GreekPsiliAndVaria,
    GreekPsiliAndOxia,
    GreekPsiliAndPerispomeni,
    GreekSmallLetterIotaWithVrachy,
    GreekSmallLetterIotaWithMacron,
    GreekSmallLetterIotaWithDialytikaAndVaria,
    GreekSmallLetterIotaWithDialytikaAndOxia,
    GreekSmallLetterIotaWithPerispomeni,
    GreekSmallLetterIotaWithDialytikaAndPerispomeni,
    GreekCapitalLetterIotaWithVrachy,
    GreekCapitalLetterIotaWithMacron,
    GreekCapitalLetterIotaWithVaria,
    GreekCapitalLetterIotaWithOxia,
    GreekDasiaAndVaria,
    GreekDasiaAndOxia,
    GreekDasiaAndPerispomeni,
    GreekSmallLetterUpsilonWithVrachy,
    GreekSmallLetterUpsilonWithMacron,
    GreekSmallLetterUpsilonWithDialytikaAndVaria,
    GreekSmallLetterUpsilonWithDialytikaAndOxia,
    GreekSmallLetterRhoWithPsili,
    GreekSmallLetterRhoWithDasia,
    GreekSmallLetterUpsilonWithPerispomeni,
    GreekSmallLetterUpsilonWithDialytikaAndPerispomeni,
    GreekCapitalLetterUpsilonWithVrachy,
    GreekCapitalLetterUpsilonWithMacron,
    GreekCapitalLetterUpsilonWithVaria,
    GreekCapitalLetterUpsilonWithOxia,
    GreekCapitalLetterRhoWithDasia,
    GreekDialytikaAndVaria,
    GreekDialytikaAndOxia,
    GreekVaria,
    GreekSmallLetterOmegaWithVariaAndYpogegrammeni,
    GreekSmallLetterOmegaWithYpogegrammeni,
    GreekSmallLetterOmegaWithOxiaAndYpogegrammeni,
    GreekSmallLetterOmegaWithPerispomeni,
    GreekSmallLetterOmegaWithPerispomeniAndYpogegrammeni,
    GreekCapitalLetterOmicronWithVaria,
    GreekCapitalLetterOmicronWithOxia,
    GreekCapitalLetterOmegaWithVaria,
    GreekCapitalLetterOmegaWithOxia,
    GreekCapitalLetterOmegaWithProsgegrammeni,
    GreekOxia,
    GreekDasia,
}

An enum to represent all characters in the GreekExtended block.

Variants
GreekSmallLetterAlphaWithPsili

\u{1f00}: 'ἀ'

GreekSmallLetterAlphaWithDasia

\u{1f01}: 'ἁ'

GreekSmallLetterAlphaWithPsiliAndVaria

\u{1f02}: 'ἂ'

GreekSmallLetterAlphaWithDasiaAndVaria

\u{1f03}: 'ἃ'

GreekSmallLetterAlphaWithPsiliAndOxia

\u{1f04}: 'ἄ'

GreekSmallLetterAlphaWithDasiaAndOxia

\u{1f05}: 'ἅ'

GreekSmallLetterAlphaWithPsiliAndPerispomeni

\u{1f06}: 'ἆ'

GreekSmallLetterAlphaWithDasiaAndPerispomeni

\u{1f07}: 'ἇ'

GreekCapitalLetterAlphaWithPsili

\u{1f08}: 'Ἀ'

GreekCapitalLetterAlphaWithDasia

\u{1f09}: 'Ἁ'

GreekCapitalLetterAlphaWithPsiliAndVaria

\u{1f0a}: 'Ἂ'

GreekCapitalLetterAlphaWithDasiaAndVaria

\u{1f0b}: 'Ἃ'

GreekCapitalLetterAlphaWithPsiliAndOxia

\u{1f0c}: 'Ἄ'

GreekCapitalLetterAlphaWithDasiaAndOxia

\u{1f0d}: 'Ἅ'

GreekCapitalLetterAlphaWithPsiliAndPerispomeni

\u{1f0e}: 'Ἆ'

GreekCapitalLetterAlphaWithDasiaAndPerispomeni

\u{1f0f}: 'Ἇ'

GreekSmallLetterEpsilonWithPsili

\u{1f10}: 'ἐ'

GreekSmallLetterEpsilonWithDasia

\u{1f11}: 'ἑ'

GreekSmallLetterEpsilonWithPsiliAndVaria

\u{1f12}: 'ἒ'

GreekSmallLetterEpsilonWithDasiaAndVaria

\u{1f13}: 'ἓ'

GreekSmallLetterEpsilonWithPsiliAndOxia

\u{1f14}: 'ἔ'

GreekSmallLetterEpsilonWithDasiaAndOxia

\u{1f15}: 'ἕ'

GreekCapitalLetterEpsilonWithPsili

\u{1f18}: 'Ἐ'

GreekCapitalLetterEpsilonWithDasia

\u{1f19}: 'Ἑ'

GreekCapitalLetterEpsilonWithPsiliAndVaria

\u{1f1a}: 'Ἒ'

GreekCapitalLetterEpsilonWithDasiaAndVaria

\u{1f1b}: 'Ἓ'

GreekCapitalLetterEpsilonWithPsiliAndOxia

\u{1f1c}: 'Ἔ'

GreekCapitalLetterEpsilonWithDasiaAndOxia

\u{1f1d}: 'Ἕ'

GreekSmallLetterEtaWithPsili

\u{1f20}: 'ἠ'

GreekSmallLetterEtaWithDasia

\u{1f21}: 'ἡ'

GreekSmallLetterEtaWithPsiliAndVaria

\u{1f22}: 'ἢ'

GreekSmallLetterEtaWithDasiaAndVaria

\u{1f23}: 'ἣ'

GreekSmallLetterEtaWithPsiliAndOxia

\u{1f24}: 'ἤ'

GreekSmallLetterEtaWithDasiaAndOxia

\u{1f25}: 'ἥ'

GreekSmallLetterEtaWithPsiliAndPerispomeni

\u{1f26}: 'ἦ'

GreekSmallLetterEtaWithDasiaAndPerispomeni

\u{1f27}: 'ἧ'

GreekCapitalLetterEtaWithPsili

\u{1f28}: 'Ἠ'

GreekCapitalLetterEtaWithDasia

\u{1f29}: 'Ἡ'

GreekCapitalLetterEtaWithPsiliAndVaria

\u{1f2a}: 'Ἢ'

GreekCapitalLetterEtaWithDasiaAndVaria

\u{1f2b}: 'Ἣ'

GreekCapitalLetterEtaWithPsiliAndOxia

\u{1f2c}: 'Ἤ'

GreekCapitalLetterEtaWithDasiaAndOxia

\u{1f2d}: 'Ἥ'

GreekCapitalLetterEtaWithPsiliAndPerispomeni

\u{1f2e}: 'Ἦ'

GreekCapitalLetterEtaWithDasiaAndPerispomeni

\u{1f2f}: 'Ἧ'

GreekSmallLetterIotaWithPsili

\u{1f30}: 'ἰ'

GreekSmallLetterIotaWithDasia

\u{1f31}: 'ἱ'

GreekSmallLetterIotaWithPsiliAndVaria

\u{1f32}: 'ἲ'

GreekSmallLetterIotaWithDasiaAndVaria

\u{1f33}: 'ἳ'

GreekSmallLetterIotaWithPsiliAndOxia

\u{1f34}: 'ἴ'

GreekSmallLetterIotaWithDasiaAndOxia

\u{1f35}: 'ἵ'

GreekSmallLetterIotaWithPsiliAndPerispomeni

\u{1f36}: 'ἶ'

GreekSmallLetterIotaWithDasiaAndPerispomeni

\u{1f37}: 'ἷ'

GreekCapitalLetterIotaWithPsili

\u{1f38}: 'Ἰ'

GreekCapitalLetterIotaWithDasia

\u{1f39}: 'Ἱ'

GreekCapitalLetterIotaWithPsiliAndVaria

\u{1f3a}: 'Ἲ'

GreekCapitalLetterIotaWithDasiaAndVaria

\u{1f3b}: 'Ἳ'

GreekCapitalLetterIotaWithPsiliAndOxia

\u{1f3c}: 'Ἴ'

GreekCapitalLetterIotaWithDasiaAndOxia

\u{1f3d}: 'Ἵ'

GreekCapitalLetterIotaWithPsiliAndPerispomeni

\u{1f3e}: 'Ἶ'

GreekCapitalLetterIotaWithDasiaAndPerispomeni

\u{1f3f}: 'Ἷ'

GreekSmallLetterOmicronWithPsili

\u{1f40}: 'ὀ'

GreekSmallLetterOmicronWithDasia

\u{1f41}: 'ὁ'

GreekSmallLetterOmicronWithPsiliAndVaria

\u{1f42}: 'ὂ'

GreekSmallLetterOmicronWithDasiaAndVaria

\u{1f43}: 'ὃ'

GreekSmallLetterOmicronWithPsiliAndOxia

\u{1f44}: 'ὄ'

GreekSmallLetterOmicronWithDasiaAndOxia

\u{1f45}: 'ὅ'

GreekCapitalLetterOmicronWithPsili

\u{1f48}: 'Ὀ'

GreekCapitalLetterOmicronWithDasia

\u{1f49}: 'Ὁ'

GreekCapitalLetterOmicronWithPsiliAndVaria

\u{1f4a}: 'Ὂ'

GreekCapitalLetterOmicronWithDasiaAndVaria

\u{1f4b}: 'Ὃ'

GreekCapitalLetterOmicronWithPsiliAndOxia

\u{1f4c}: 'Ὄ'

GreekCapitalLetterOmicronWithDasiaAndOxia

\u{1f4d}: 'Ὅ'

GreekSmallLetterUpsilonWithPsili

\u{1f50}: 'ὐ'

GreekSmallLetterUpsilonWithDasia

\u{1f51}: 'ὑ'

GreekSmallLetterUpsilonWithPsiliAndVaria

\u{1f52}: 'ὒ'

GreekSmallLetterUpsilonWithDasiaAndVaria

\u{1f53}: 'ὓ'

GreekSmallLetterUpsilonWithPsiliAndOxia

\u{1f54}: 'ὔ'

GreekSmallLetterUpsilonWithDasiaAndOxia

\u{1f55}: 'ὕ'

GreekSmallLetterUpsilonWithPsiliAndPerispomeni

\u{1f56}: 'ὖ'

GreekSmallLetterUpsilonWithDasiaAndPerispomeni

\u{1f57}: 'ὗ'

GreekCapitalLetterUpsilonWithDasia

\u{1f59}: 'Ὑ'

GreekCapitalLetterUpsilonWithDasiaAndVaria

\u{1f5b}: 'Ὓ'

GreekCapitalLetterUpsilonWithDasiaAndOxia

\u{1f5d}: 'Ὕ'

GreekCapitalLetterUpsilonWithDasiaAndPerispomeni

\u{1f5f}: 'Ὗ'

GreekSmallLetterOmegaWithPsili

\u{1f60}: 'ὠ'

GreekSmallLetterOmegaWithDasia

\u{1f61}: 'ὡ'

GreekSmallLetterOmegaWithPsiliAndVaria

\u{1f62}: 'ὢ'

GreekSmallLetterOmegaWithDasiaAndVaria

\u{1f63}: 'ὣ'

GreekSmallLetterOmegaWithPsiliAndOxia

\u{1f64}: 'ὤ'

GreekSmallLetterOmegaWithDasiaAndOxia

\u{1f65}: 'ὥ'

GreekSmallLetterOmegaWithPsiliAndPerispomeni

\u{1f66}: 'ὦ'

GreekSmallLetterOmegaWithDasiaAndPerispomeni

\u{1f67}: 'ὧ'

GreekCapitalLetterOmegaWithPsili

\u{1f68}: 'Ὠ'

GreekCapitalLetterOmegaWithDasia

\u{1f69}: 'Ὡ'

GreekCapitalLetterOmegaWithPsiliAndVaria

\u{1f6a}: 'Ὢ'

GreekCapitalLetterOmegaWithDasiaAndVaria

\u{1f6b}: 'Ὣ'

GreekCapitalLetterOmegaWithPsiliAndOxia

\u{1f6c}: 'Ὤ'

GreekCapitalLetterOmegaWithDasiaAndOxia

\u{1f6d}: 'Ὥ'

GreekCapitalLetterOmegaWithPsiliAndPerispomeni

\u{1f6e}: 'Ὦ'

GreekCapitalLetterOmegaWithDasiaAndPerispomeni

\u{1f6f}: 'Ὧ'

GreekSmallLetterAlphaWithVaria

\u{1f70}: 'ὰ'

GreekSmallLetterAlphaWithOxia

\u{1f71}: 'ά'

GreekSmallLetterEpsilonWithVaria

\u{1f72}: 'ὲ'

GreekSmallLetterEpsilonWithOxia

\u{1f73}: 'έ'

GreekSmallLetterEtaWithVaria

\u{1f74}: 'ὴ'

GreekSmallLetterEtaWithOxia

\u{1f75}: 'ή'

GreekSmallLetterIotaWithVaria

\u{1f76}: 'ὶ'

GreekSmallLetterIotaWithOxia

\u{1f77}: 'ί'

GreekSmallLetterOmicronWithVaria

\u{1f78}: 'ὸ'

GreekSmallLetterOmicronWithOxia

\u{1f79}: 'ό'

GreekSmallLetterUpsilonWithVaria

\u{1f7a}: 'ὺ'

GreekSmallLetterUpsilonWithOxia

\u{1f7b}: 'ύ'

GreekSmallLetterOmegaWithVaria

\u{1f7c}: 'ὼ'

GreekSmallLetterOmegaWithOxia

\u{1f7d}: 'ώ'

GreekSmallLetterAlphaWithPsiliAndYpogegrammeni

\u{1f80}: 'ᾀ'

GreekSmallLetterAlphaWithDasiaAndYpogegrammeni

\u{1f81}: 'ᾁ'

GreekSmallLetterAlphaWithPsiliAndVariaAndYpogegrammeni

\u{1f82}: 'ᾂ'

GreekSmallLetterAlphaWithDasiaAndVariaAndYpogegrammeni

\u{1f83}: 'ᾃ'

GreekSmallLetterAlphaWithPsiliAndOxiaAndYpogegrammeni

\u{1f84}: 'ᾄ'

GreekSmallLetterAlphaWithDasiaAndOxiaAndYpogegrammeni

\u{1f85}: 'ᾅ'

GreekSmallLetterAlphaWithPsiliAndPerispomeniAndYpogegrammeni

\u{1f86}: 'ᾆ'

GreekSmallLetterAlphaWithDasiaAndPerispomeniAndYpogegrammeni

\u{1f87}: 'ᾇ'

GreekCapitalLetterAlphaWithPsiliAndProsgegrammeni

\u{1f88}: 'ᾈ'

GreekCapitalLetterAlphaWithDasiaAndProsgegrammeni

\u{1f89}: 'ᾉ'

GreekCapitalLetterAlphaWithPsiliAndVariaAndProsgegrammeni

\u{1f8a}: 'ᾊ'

GreekCapitalLetterAlphaWithDasiaAndVariaAndProsgegrammeni

\u{1f8b}: 'ᾋ'

GreekCapitalLetterAlphaWithPsiliAndOxiaAndProsgegrammeni

\u{1f8c}: 'ᾌ'

GreekCapitalLetterAlphaWithDasiaAndOxiaAndProsgegrammeni

\u{1f8d}: 'ᾍ'

GreekCapitalLetterAlphaWithPsiliAndPerispomeniAndProsgegrammeni

\u{1f8e}: 'ᾎ'

GreekCapitalLetterAlphaWithDasiaAndPerispomeniAndProsgegrammeni

\u{1f8f}: 'ᾏ'

GreekSmallLetterEtaWithPsiliAndYpogegrammeni

\u{1f90}: 'ᾐ'

GreekSmallLetterEtaWithDasiaAndYpogegrammeni

\u{1f91}: 'ᾑ'

GreekSmallLetterEtaWithPsiliAndVariaAndYpogegrammeni

\u{1f92}: 'ᾒ'

GreekSmallLetterEtaWithDasiaAndVariaAndYpogegrammeni

\u{1f93}: 'ᾓ'

GreekSmallLetterEtaWithPsiliAndOxiaAndYpogegrammeni

\u{1f94}: 'ᾔ'

GreekSmallLetterEtaWithDasiaAndOxiaAndYpogegrammeni

\u{1f95}: 'ᾕ'

GreekSmallLetterEtaWithPsiliAndPerispomeniAndYpogegrammeni

\u{1f96}: 'ᾖ'

GreekSmallLetterEtaWithDasiaAndPerispomeniAndYpogegrammeni

\u{1f97}: 'ᾗ'

GreekCapitalLetterEtaWithPsiliAndProsgegrammeni

\u{1f98}: 'ᾘ'

GreekCapitalLetterEtaWithDasiaAndProsgegrammeni

\u{1f99}: 'ᾙ'

GreekCapitalLetterEtaWithPsiliAndVariaAndProsgegrammeni

\u{1f9a}: 'ᾚ'

GreekCapitalLetterEtaWithDasiaAndVariaAndProsgegrammeni

\u{1f9b}: 'ᾛ'

GreekCapitalLetterEtaWithPsiliAndOxiaAndProsgegrammeni

\u{1f9c}: 'ᾜ'

GreekCapitalLetterEtaWithDasiaAndOxiaAndProsgegrammeni

\u{1f9d}: 'ᾝ'

GreekCapitalLetterEtaWithPsiliAndPerispomeniAndProsgegrammeni

\u{1f9e}: 'ᾞ'

GreekCapitalLetterEtaWithDasiaAndPerispomeniAndProsgegrammeni

\u{1f9f}: 'ᾟ'

GreekSmallLetterOmegaWithPsiliAndYpogegrammeni

\u{1fa0}: 'ᾠ'

GreekSmallLetterOmegaWithDasiaAndYpogegrammeni

\u{1fa1}: 'ᾡ'

GreekSmallLetterOmegaWithPsiliAndVariaAndYpogegrammeni

\u{1fa2}: 'ᾢ'

GreekSmallLetterOmegaWithDasiaAndVariaAndYpogegrammeni

\u{1fa3}: 'ᾣ'

GreekSmallLetterOmegaWithPsiliAndOxiaAndYpogegrammeni

\u{1fa4}: 'ᾤ'

GreekSmallLetterOmegaWithDasiaAndOxiaAndYpogegrammeni

\u{1fa5}: 'ᾥ'

GreekSmallLetterOmegaWithPsiliAndPerispomeniAndYpogegrammeni

\u{1fa6}: 'ᾦ'

GreekSmallLetterOmegaWithDasiaAndPerispomeniAndYpogegrammeni

\u{1fa7}: 'ᾧ'

GreekCapitalLetterOmegaWithPsiliAndProsgegrammeni

\u{1fa8}: 'ᾨ'

GreekCapitalLetterOmegaWithDasiaAndProsgegrammeni

\u{1fa9}: 'ᾩ'

GreekCapitalLetterOmegaWithPsiliAndVariaAndProsgegrammeni

\u{1faa}: 'ᾪ'

GreekCapitalLetterOmegaWithDasiaAndVariaAndProsgegrammeni

\u{1fab}: 'ᾫ'

GreekCapitalLetterOmegaWithPsiliAndOxiaAndProsgegrammeni

\u{1fac}: 'ᾬ'

GreekCapitalLetterOmegaWithDasiaAndOxiaAndProsgegrammeni

\u{1fad}: 'ᾭ'

GreekCapitalLetterOmegaWithPsiliAndPerispomeniAndProsgegrammeni

\u{1fae}: 'ᾮ'

GreekCapitalLetterOmegaWithDasiaAndPerispomeniAndProsgegrammeni

\u{1faf}: 'ᾯ'

GreekSmallLetterAlphaWithVrachy

\u{1fb0}: 'ᾰ'

GreekSmallLetterAlphaWithMacron

\u{1fb1}: 'ᾱ'

GreekSmallLetterAlphaWithVariaAndYpogegrammeni

\u{1fb2}: 'ᾲ'

GreekSmallLetterAlphaWithYpogegrammeni

\u{1fb3}: 'ᾳ'

GreekSmallLetterAlphaWithOxiaAndYpogegrammeni

\u{1fb4}: 'ᾴ'

GreekSmallLetterAlphaWithPerispomeni

\u{1fb6}: 'ᾶ'

GreekSmallLetterAlphaWithPerispomeniAndYpogegrammeni

\u{1fb7}: 'ᾷ'

GreekCapitalLetterAlphaWithVrachy

\u{1fb8}: 'Ᾰ'

GreekCapitalLetterAlphaWithMacron

\u{1fb9}: 'Ᾱ'

GreekCapitalLetterAlphaWithVaria

\u{1fba}: 'Ὰ'

GreekCapitalLetterAlphaWithOxia

\u{1fbb}: 'Ά'

GreekCapitalLetterAlphaWithProsgegrammeni

\u{1fbc}: 'ᾼ'

GreekKoronis

\u{1fbd}: '᾽'

GreekProsgegrammeni

\u{1fbe}: 'ι'

GreekPsili

\u{1fbf}: '᾿'

GreekPerispomeni

\u{1fc0}: '῀'

GreekDialytikaAndPerispomeni

\u{1fc1}: '῁'

GreekSmallLetterEtaWithVariaAndYpogegrammeni

\u{1fc2}: 'ῂ'

GreekSmallLetterEtaWithYpogegrammeni

\u{1fc3}: 'ῃ'

GreekSmallLetterEtaWithOxiaAndYpogegrammeni

\u{1fc4}: 'ῄ'

GreekSmallLetterEtaWithPerispomeni

\u{1fc6}: 'ῆ'

GreekSmallLetterEtaWithPerispomeniAndYpogegrammeni

\u{1fc7}: 'ῇ'

GreekCapitalLetterEpsilonWithVaria

\u{1fc8}: 'Ὲ'

GreekCapitalLetterEpsilonWithOxia

\u{1fc9}: 'Έ'

GreekCapitalLetterEtaWithVaria

\u{1fca}: 'Ὴ'

GreekCapitalLetterEtaWithOxia

\u{1fcb}: 'Ή'

GreekCapitalLetterEtaWithProsgegrammeni

\u{1fcc}: 'ῌ'

GreekPsiliAndVaria

\u{1fcd}: '῍'

GreekPsiliAndOxia

\u{1fce}: '῎'

GreekPsiliAndPerispomeni

\u{1fcf}: '῏'

GreekSmallLetterIotaWithVrachy

\u{1fd0}: 'ῐ'

GreekSmallLetterIotaWithMacron

\u{1fd1}: 'ῑ'

GreekSmallLetterIotaWithDialytikaAndVaria

\u{1fd2}: 'ῒ'

GreekSmallLetterIotaWithDialytikaAndOxia

\u{1fd3}: 'ΐ'

GreekSmallLetterIotaWithPerispomeni

\u{1fd6}: 'ῖ'

GreekSmallLetterIotaWithDialytikaAndPerispomeni

\u{1fd7}: 'ῗ'

GreekCapitalLetterIotaWithVrachy

\u{1fd8}: 'Ῐ'

GreekCapitalLetterIotaWithMacron

\u{1fd9}: 'Ῑ'

GreekCapitalLetterIotaWithVaria

\u{1fda}: 'Ὶ'

GreekCapitalLetterIotaWithOxia

\u{1fdb}: 'Ί'

GreekDasiaAndVaria

\u{1fdd}: '῝'

GreekDasiaAndOxia

\u{1fde}: '῞'

GreekDasiaAndPerispomeni

\u{1fdf}: '῟'

GreekSmallLetterUpsilonWithVrachy

\u{1fe0}: 'ῠ'

GreekSmallLetterUpsilonWithMacron

\u{1fe1}: 'ῡ'

GreekSmallLetterUpsilonWithDialytikaAndVaria

\u{1fe2}: 'ῢ'

GreekSmallLetterUpsilonWithDialytikaAndOxia

\u{1fe3}: 'ΰ'

GreekSmallLetterRhoWithPsili

\u{1fe4}: 'ῤ'

GreekSmallLetterRhoWithDasia

\u{1fe5}: 'ῥ'

GreekSmallLetterUpsilonWithPerispomeni

\u{1fe6}: 'ῦ'

GreekSmallLetterUpsilonWithDialytikaAndPerispomeni

\u{1fe7}: 'ῧ'

GreekCapitalLetterUpsilonWithVrachy

\u{1fe8}: 'Ῠ'

GreekCapitalLetterUpsilonWithMacron

\u{1fe9}: 'Ῡ'

GreekCapitalLetterUpsilonWithVaria

\u{1fea}: 'Ὺ'

GreekCapitalLetterUpsilonWithOxia

\u{1feb}: 'Ύ'

GreekCapitalLetterRhoWithDasia

\u{1fec}: 'Ῥ'

GreekDialytikaAndVaria

\u{1fed}: '῭'

GreekDialytikaAndOxia

\u{1fee}: '΅'

GreekVaria

\u{1fef}: '`'

GreekSmallLetterOmegaWithVariaAndYpogegrammeni

\u{1ff2}: 'ῲ'

GreekSmallLetterOmegaWithYpogegrammeni

\u{1ff3}: 'ῳ'

GreekSmallLetterOmegaWithOxiaAndYpogegrammeni

\u{1ff4}: 'ῴ'

GreekSmallLetterOmegaWithPerispomeni

\u{1ff6}: 'ῶ'

GreekSmallLetterOmegaWithPerispomeniAndYpogegrammeni

\u{1ff7}: 'ῷ'

GreekCapitalLetterOmicronWithVaria

\u{1ff8}: 'Ὸ'

GreekCapitalLetterOmicronWithOxia

\u{1ff9}: 'Ό'

GreekCapitalLetterOmegaWithVaria

\u{1ffa}: 'Ὼ'

GreekCapitalLetterOmegaWithOxia

\u{1ffb}: 'Ώ'

GreekCapitalLetterOmegaWithProsgegrammeni

\u{1ffc}: 'ῼ'

GreekOxia

\u{1ffd}: '´'

GreekDasia

\u{1ffe}: '῾'

Methods

impl GreekExtended

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case