Enum unicode_types::halfwidth_and_fullwidth_forms::HalfwidthandFullwidthForms

pub enum HalfwidthandFullwidthForms {
    FullwidthExclamationMark,
    FullwidthQuotationMark,
    FullwidthNumberSign,
    FullwidthDollarSign,
    FullwidthPercentSign,
    FullwidthAmpersand,
    FullwidthApostrophe,
    FullwidthLeftParenthesis,
    FullwidthRightParenthesis,
    FullwidthAsterisk,
    FullwidthPlusSign,
    FullwidthComma,
    FullwidthHyphenDashMinus,
    FullwidthFullStop,
    FullwidthSolidus,
    FullwidthDigitZero,
    FullwidthDigitOne,
    FullwidthDigitTwo,
    FullwidthDigitThree,
    FullwidthDigitFour,
    FullwidthDigitFive,
    FullwidthDigitSix,
    FullwidthDigitSeven,
    FullwidthDigitEight,
    FullwidthDigitNine,
    FullwidthColon,
    FullwidthSemicolon,
    FullwidthLessDashThanSign,
    FullwidthEqualsSign,
    FullwidthGreaterDashThanSign,
    FullwidthQuestionMark,
    FullwidthCommercialAt,
    FullwidthLatinCapitalLetterA,
    FullwidthLatinCapitalLetterB,
    FullwidthLatinCapitalLetterC,
    FullwidthLatinCapitalLetterD,
    FullwidthLatinCapitalLetterE,
    FullwidthLatinCapitalLetterF,
    FullwidthLatinCapitalLetterG,
    FullwidthLatinCapitalLetterH,
    FullwidthLatinCapitalLetterI,
    FullwidthLatinCapitalLetterJ,
    FullwidthLatinCapitalLetterK,
    FullwidthLatinCapitalLetterL,
    FullwidthLatinCapitalLetterM,
    FullwidthLatinCapitalLetterN,
    FullwidthLatinCapitalLetterO,
    FullwidthLatinCapitalLetterP,
    FullwidthLatinCapitalLetterQ,
    FullwidthLatinCapitalLetterR,
    FullwidthLatinCapitalLetterS,
    FullwidthLatinCapitalLetterT,
    FullwidthLatinCapitalLetterU,
    FullwidthLatinCapitalLetterV,
    FullwidthLatinCapitalLetterW,
    FullwidthLatinCapitalLetterX,
    FullwidthLatinCapitalLetterY,
    FullwidthLatinCapitalLetterZ,
    FullwidthLeftSquareBracket,
    FullwidthReverseSolidus,
    FullwidthRightSquareBracket,
    FullwidthCircumflexAccent,
    FullwidthLowLine,
    FullwidthGraveAccent,
    FullwidthLatinSmallLetterA,
    FullwidthLatinSmallLetterB,
    FullwidthLatinSmallLetterC,
    FullwidthLatinSmallLetterD,
    FullwidthLatinSmallLetterE,
    FullwidthLatinSmallLetterF,
    FullwidthLatinSmallLetterG,
    FullwidthLatinSmallLetterH,
    FullwidthLatinSmallLetterI,
    FullwidthLatinSmallLetterJ,
    FullwidthLatinSmallLetterK,
    FullwidthLatinSmallLetterL,
    FullwidthLatinSmallLetterM,
    FullwidthLatinSmallLetterN,
    FullwidthLatinSmallLetterO,
    FullwidthLatinSmallLetterP,
    FullwidthLatinSmallLetterQ,
    FullwidthLatinSmallLetterR,
    FullwidthLatinSmallLetterS,
    FullwidthLatinSmallLetterT,
    FullwidthLatinSmallLetterU,
    FullwidthLatinSmallLetterV,
    FullwidthLatinSmallLetterW,
    FullwidthLatinSmallLetterX,
    FullwidthLatinSmallLetterY,
    FullwidthLatinSmallLetterZ,
    FullwidthLeftCurlyBracket,
    FullwidthVerticalLine,
    FullwidthRightCurlyBracket,
    FullwidthTilde,
    FullwidthLeftWhiteParenthesis,
    FullwidthRightWhiteParenthesis,
    HalfwidthIdeographicFullStop,
    HalfwidthLeftCornerBracket,
    HalfwidthRightCornerBracket,
    HalfwidthIdeographicComma,
    HalfwidthKatakanaMiddleDot,
    HalfwidthKatakanaLetterWo,
    HalfwidthKatakanaLetterSmallA,
    HalfwidthKatakanaLetterSmallI,
    HalfwidthKatakanaLetterSmallU,
    HalfwidthKatakanaLetterSmallE,
    HalfwidthKatakanaLetterSmallO,
    HalfwidthKatakanaLetterSmallYa,
    HalfwidthKatakanaLetterSmallYu,
    HalfwidthKatakanaLetterSmallYo,
    HalfwidthKatakanaLetterSmallTu,
    HalfwidthKatakanaDashHiraganaProlongedSoundMark,
    HalfwidthKatakanaLetterA,
    HalfwidthKatakanaLetterI,
    HalfwidthKatakanaLetterU,
    HalfwidthKatakanaLetterE,
    HalfwidthKatakanaLetterO,
    HalfwidthKatakanaLetterKa,
    HalfwidthKatakanaLetterKi,
    HalfwidthKatakanaLetterKu,
    HalfwidthKatakanaLetterKe,
    HalfwidthKatakanaLetterKo,
    HalfwidthKatakanaLetterSa,
    HalfwidthKatakanaLetterSi,
    HalfwidthKatakanaLetterSu,
    HalfwidthKatakanaLetterSe,
    HalfwidthKatakanaLetterSo,
    HalfwidthKatakanaLetterTa,
    HalfwidthKatakanaLetterTi,
    HalfwidthKatakanaLetterTu,
    HalfwidthKatakanaLetterTe,
    HalfwidthKatakanaLetterTo,
    HalfwidthKatakanaLetterNa,
    HalfwidthKatakanaLetterNi,
    HalfwidthKatakanaLetterNu,
    HalfwidthKatakanaLetterNe,
    HalfwidthKatakanaLetterNo,
    HalfwidthKatakanaLetterHa,
    HalfwidthKatakanaLetterHi,
    HalfwidthKatakanaLetterHu,
    HalfwidthKatakanaLetterHe,
    HalfwidthKatakanaLetterHo,
    HalfwidthKatakanaLetterMa,
    HalfwidthKatakanaLetterMi,
    HalfwidthKatakanaLetterMu,
    HalfwidthKatakanaLetterMe,
    HalfwidthKatakanaLetterMo,
    HalfwidthKatakanaLetterYa,
    HalfwidthKatakanaLetterYu,
    HalfwidthKatakanaLetterYo,
    HalfwidthKatakanaLetterRa,
    HalfwidthKatakanaLetterRi,
    HalfwidthKatakanaLetterRu,
    HalfwidthKatakanaLetterRe,
    HalfwidthKatakanaLetterRo,
    HalfwidthKatakanaLetterWa,
    HalfwidthKatakanaLetterN,
    HalfwidthKatakanaVoicedSoundMark,
    HalfwidthKatakanaSemiDashVoicedSoundMark,
    HalfwidthHangulFiller,
    HalfwidthHangulLetterKiyeok,
    HalfwidthHangulLetterSsangkiyeok,
    HalfwidthHangulLetterKiyeokDashSios,
    HalfwidthHangulLetterNieun,
    HalfwidthHangulLetterNieunDashCieuc,
    HalfwidthHangulLetterNieunDashHieuh,
    HalfwidthHangulLetterTikeut,
    HalfwidthHangulLetterSsangtikeut,
    HalfwidthHangulLetterRieul,
    HalfwidthHangulLetterRieulDashKiyeok,
    HalfwidthHangulLetterRieulDashMieum,
    HalfwidthHangulLetterRieulDashPieup,
    HalfwidthHangulLetterRieulDashSios,
    HalfwidthHangulLetterRieulDashThieuth,
    HalfwidthHangulLetterRieulDashPhieuph,
    HalfwidthHangulLetterRieulDashHieuh,
    HalfwidthHangulLetterMieum,
    HalfwidthHangulLetterPieup,
    HalfwidthHangulLetterSsangpieup,
    HalfwidthHangulLetterPieupDashSios,
    HalfwidthHangulLetterSios,
    HalfwidthHangulLetterSsangsios,
    HalfwidthHangulLetterIeung,
    HalfwidthHangulLetterCieuc,
    HalfwidthHangulLetterSsangcieuc,
    HalfwidthHangulLetterChieuch,
    HalfwidthHangulLetterKhieukh,
    HalfwidthHangulLetterThieuth,
    HalfwidthHangulLetterPhieuph,
    HalfwidthHangulLetterHieuh,
    HalfwidthHangulLetterA,
    HalfwidthHangulLetterAe,
    HalfwidthHangulLetterYa,
    HalfwidthHangulLetterYae,
    HalfwidthHangulLetterEo,
    HalfwidthHangulLetterE,
    HalfwidthHangulLetterYeo,
    HalfwidthHangulLetterYe,
    HalfwidthHangulLetterO,
    HalfwidthHangulLetterWa,
    HalfwidthHangulLetterWae,
    HalfwidthHangulLetterOe,
    HalfwidthHangulLetterYo,
    HalfwidthHangulLetterU,
    HalfwidthHangulLetterWeo,
    HalfwidthHangulLetterWe,
    HalfwidthHangulLetterWi,
    HalfwidthHangulLetterYu,
    HalfwidthHangulLetterEu,
    HalfwidthHangulLetterYi,
    HalfwidthHangulLetterI,
    FullwidthCentSign,
    FullwidthPoundSign,
    FullwidthNotSign,
    FullwidthMacron,
    FullwidthBrokenBar,
    FullwidthYenSign,
    FullwidthWonSign,
    HalfwidthFormsLightVertical,
    HalfwidthLeftwardsArrow,
    HalfwidthUpwardsArrow,
    HalfwidthRightwardsArrow,
    HalfwidthDownwardsArrow,
    HalfwidthBlackSquare,
    HalfwidthWhiteCircle,
}

An enum to represent all characters in the HalfwidthandFullwidthForms block.

Variants
FullwidthExclamationMark

\u{ff01}: '！'

FullwidthQuotationMark

\u{ff02}: '＂'

FullwidthNumberSign

\u{ff03}: '＃'

FullwidthDollarSign

\u{ff04}: '＄'

FullwidthPercentSign

\u{ff05}: '％'

FullwidthAmpersand

\u{ff06}: '＆'

FullwidthApostrophe

\u{ff07}: '＇'

FullwidthLeftParenthesis

\u{ff08}: '（'

FullwidthRightParenthesis

\u{ff09}: '）'

FullwidthAsterisk

\u{ff0a}: '＊'

FullwidthPlusSign

\u{ff0b}: '＋'

FullwidthComma

\u{ff0c}: '，'

FullwidthHyphenDashMinus

\u{ff0d}: '－'

FullwidthFullStop

\u{ff0e}: '．'

FullwidthSolidus

\u{ff0f}: '／'

FullwidthDigitZero

\u{ff10}: '０'

FullwidthDigitOne

\u{ff11}: '１'

FullwidthDigitTwo

\u{ff12}: '２'

FullwidthDigitThree

\u{ff13}: '３'

FullwidthDigitFour

\u{ff14}: '４'

FullwidthDigitFive

\u{ff15}: '５'

FullwidthDigitSix

\u{ff16}: '６'

FullwidthDigitSeven

\u{ff17}: '７'

FullwidthDigitEight

\u{ff18}: '８'

FullwidthDigitNine

\u{ff19}: '９'

FullwidthColon

\u{ff1a}: '：'

FullwidthSemicolon

\u{ff1b}: '；'

FullwidthLessDashThanSign

\u{ff1c}: '＜'

FullwidthEqualsSign

\u{ff1d}: '＝'

FullwidthGreaterDashThanSign

\u{ff1e}: '＞'

FullwidthQuestionMark

\u{ff1f}: '？'

FullwidthCommercialAt

\u{ff20}: '＠'

FullwidthLatinCapitalLetterA

\u{ff21}: 'Ａ'

FullwidthLatinCapitalLetterB

\u{ff22}: 'Ｂ'

FullwidthLatinCapitalLetterC

\u{ff23}: 'Ｃ'

FullwidthLatinCapitalLetterD

\u{ff24}: 'Ｄ'

FullwidthLatinCapitalLetterE

\u{ff25}: 'Ｅ'

FullwidthLatinCapitalLetterF

\u{ff26}: 'Ｆ'

FullwidthLatinCapitalLetterG

\u{ff27}: 'Ｇ'

FullwidthLatinCapitalLetterH

\u{ff28}: 'Ｈ'

FullwidthLatinCapitalLetterI

\u{ff29}: 'Ｉ'

FullwidthLatinCapitalLetterJ

\u{ff2a}: 'Ｊ'

FullwidthLatinCapitalLetterK

\u{ff2b}: 'Ｋ'

FullwidthLatinCapitalLetterL

\u{ff2c}: 'Ｌ'

FullwidthLatinCapitalLetterM

\u{ff2d}: 'Ｍ'

FullwidthLatinCapitalLetterN

\u{ff2e}: 'Ｎ'

FullwidthLatinCapitalLetterO

\u{ff2f}: 'Ｏ'

FullwidthLatinCapitalLetterP

\u{ff30}: 'Ｐ'

FullwidthLatinCapitalLetterQ

\u{ff31}: 'Ｑ'

FullwidthLatinCapitalLetterR

\u{ff32}: 'Ｒ'

FullwidthLatinCapitalLetterS

\u{ff33}: 'Ｓ'

FullwidthLatinCapitalLetterT

\u{ff34}: 'Ｔ'

FullwidthLatinCapitalLetterU

\u{ff35}: 'Ｕ'

FullwidthLatinCapitalLetterV

\u{ff36}: 'Ｖ'

FullwidthLatinCapitalLetterW

\u{ff37}: 'Ｗ'

FullwidthLatinCapitalLetterX

\u{ff38}: 'Ｘ'

FullwidthLatinCapitalLetterY

\u{ff39}: 'Ｙ'

FullwidthLatinCapitalLetterZ

\u{ff3a}: 'Ｚ'

FullwidthLeftSquareBracket

\u{ff3b}: '［'

FullwidthReverseSolidus

\u{ff3c}: '＼'

FullwidthRightSquareBracket

\u{ff3d}: '］'

FullwidthCircumflexAccent

\u{ff3e}: '＾'

FullwidthLowLine

\u{ff3f}: '＿'

FullwidthGraveAccent

\u{ff40}: '｀'

FullwidthLatinSmallLetterA

\u{ff41}: 'ａ'

FullwidthLatinSmallLetterB

\u{ff42}: 'ｂ'

FullwidthLatinSmallLetterC

\u{ff43}: 'ｃ'

FullwidthLatinSmallLetterD

\u{ff44}: 'ｄ'

FullwidthLatinSmallLetterE

\u{ff45}: 'ｅ'

FullwidthLatinSmallLetterF

\u{ff46}: 'ｆ'

FullwidthLatinSmallLetterG

\u{ff47}: 'ｇ'

FullwidthLatinSmallLetterH

\u{ff48}: 'ｈ'

FullwidthLatinSmallLetterI

\u{ff49}: 'ｉ'

FullwidthLatinSmallLetterJ

\u{ff4a}: 'ｊ'

FullwidthLatinSmallLetterK

\u{ff4b}: 'ｋ'

FullwidthLatinSmallLetterL

\u{ff4c}: 'ｌ'

FullwidthLatinSmallLetterM

\u{ff4d}: 'ｍ'

FullwidthLatinSmallLetterN

\u{ff4e}: 'ｎ'

FullwidthLatinSmallLetterO

\u{ff4f}: 'ｏ'

FullwidthLatinSmallLetterP

\u{ff50}: 'ｐ'

FullwidthLatinSmallLetterQ

\u{ff51}: 'ｑ'

FullwidthLatinSmallLetterR

\u{ff52}: 'ｒ'

FullwidthLatinSmallLetterS

\u{ff53}: 'ｓ'

FullwidthLatinSmallLetterT

\u{ff54}: 'ｔ'

FullwidthLatinSmallLetterU

\u{ff55}: 'ｕ'

FullwidthLatinSmallLetterV

\u{ff56}: 'ｖ'

FullwidthLatinSmallLetterW

\u{ff57}: 'ｗ'

FullwidthLatinSmallLetterX

\u{ff58}: 'ｘ'

FullwidthLatinSmallLetterY

\u{ff59}: 'ｙ'

FullwidthLatinSmallLetterZ

\u{ff5a}: 'ｚ'

FullwidthLeftCurlyBracket

\u{ff5b}: '｛'

FullwidthVerticalLine

\u{ff5c}: '｜'

FullwidthRightCurlyBracket

\u{ff5d}: '｝'

FullwidthTilde

\u{ff5e}: '～'

FullwidthLeftWhiteParenthesis

\u{ff5f}: '｟'

FullwidthRightWhiteParenthesis

\u{ff60}: '｠'

HalfwidthIdeographicFullStop

\u{ff61}: '｡'

HalfwidthLeftCornerBracket

\u{ff62}: '｢'

HalfwidthRightCornerBracket

\u{ff63}: '｣'

HalfwidthIdeographicComma

\u{ff64}: '､'

HalfwidthKatakanaMiddleDot

\u{ff65}: '･'

HalfwidthKatakanaLetterWo

\u{ff66}: 'ｦ'

HalfwidthKatakanaLetterSmallA

\u{ff67}: 'ｧ'

HalfwidthKatakanaLetterSmallI

\u{ff68}: 'ｨ'

HalfwidthKatakanaLetterSmallU

\u{ff69}: 'ｩ'

HalfwidthKatakanaLetterSmallE

\u{ff6a}: 'ｪ'

HalfwidthKatakanaLetterSmallO

\u{ff6b}: 'ｫ'

HalfwidthKatakanaLetterSmallYa

\u{ff6c}: 'ｬ'

HalfwidthKatakanaLetterSmallYu

\u{ff6d}: 'ｭ'

HalfwidthKatakanaLetterSmallYo

\u{ff6e}: 'ｮ'

HalfwidthKatakanaLetterSmallTu

\u{ff6f}: 'ｯ'

HalfwidthKatakanaDashHiraganaProlongedSoundMark

\u{ff70}: 'ｰ'

HalfwidthKatakanaLetterA

\u{ff71}: 'ｱ'

HalfwidthKatakanaLetterI

\u{ff72}: 'ｲ'

HalfwidthKatakanaLetterU

\u{ff73}: 'ｳ'

HalfwidthKatakanaLetterE

\u{ff74}: 'ｴ'

HalfwidthKatakanaLetterO

\u{ff75}: 'ｵ'

HalfwidthKatakanaLetterKa

\u{ff76}: 'ｶ'

HalfwidthKatakanaLetterKi

\u{ff77}: 'ｷ'

HalfwidthKatakanaLetterKu

\u{ff78}: 'ｸ'

HalfwidthKatakanaLetterKe

\u{ff79}: 'ｹ'

HalfwidthKatakanaLetterKo

\u{ff7a}: 'ｺ'

HalfwidthKatakanaLetterSa

\u{ff7b}: 'ｻ'

HalfwidthKatakanaLetterSi

\u{ff7c}: 'ｼ'

HalfwidthKatakanaLetterSu

\u{ff7d}: 'ｽ'

HalfwidthKatakanaLetterSe

\u{ff7e}: 'ｾ'

HalfwidthKatakanaLetterSo

\u{ff7f}: 'ｿ'

HalfwidthKatakanaLetterTa

\u{ff80}: 'ﾀ'

HalfwidthKatakanaLetterTi

\u{ff81}: 'ﾁ'

HalfwidthKatakanaLetterTu

\u{ff82}: 'ﾂ'

HalfwidthKatakanaLetterTe

\u{ff83}: 'ﾃ'

HalfwidthKatakanaLetterTo

\u{ff84}: 'ﾄ'

HalfwidthKatakanaLetterNa

\u{ff85}: 'ﾅ'

HalfwidthKatakanaLetterNi

\u{ff86}: 'ﾆ'

HalfwidthKatakanaLetterNu

\u{ff87}: 'ﾇ'

HalfwidthKatakanaLetterNe

\u{ff88}: 'ﾈ'

HalfwidthKatakanaLetterNo

\u{ff89}: 'ﾉ'

HalfwidthKatakanaLetterHa

\u{ff8a}: 'ﾊ'

HalfwidthKatakanaLetterHi

\u{ff8b}: 'ﾋ'

HalfwidthKatakanaLetterHu

\u{ff8c}: 'ﾌ'

HalfwidthKatakanaLetterHe

\u{ff8d}: 'ﾍ'

HalfwidthKatakanaLetterHo

\u{ff8e}: 'ﾎ'

HalfwidthKatakanaLetterMa

\u{ff8f}: 'ﾏ'

HalfwidthKatakanaLetterMi

\u{ff90}: 'ﾐ'

HalfwidthKatakanaLetterMu

\u{ff91}: 'ﾑ'

HalfwidthKatakanaLetterMe

\u{ff92}: 'ﾒ'

HalfwidthKatakanaLetterMo

\u{ff93}: 'ﾓ'

HalfwidthKatakanaLetterYa

\u{ff94}: 'ﾔ'

HalfwidthKatakanaLetterYu

\u{ff95}: 'ﾕ'

HalfwidthKatakanaLetterYo

\u{ff96}: 'ﾖ'

HalfwidthKatakanaLetterRa

\u{ff97}: 'ﾗ'

HalfwidthKatakanaLetterRi

\u{ff98}: 'ﾘ'

HalfwidthKatakanaLetterRu

\u{ff99}: 'ﾙ'

HalfwidthKatakanaLetterRe

\u{ff9a}: 'ﾚ'

HalfwidthKatakanaLetterRo

\u{ff9b}: 'ﾛ'

HalfwidthKatakanaLetterWa

\u{ff9c}: 'ﾜ'

HalfwidthKatakanaLetterN

\u{ff9d}: 'ﾝ'

HalfwidthKatakanaVoicedSoundMark

\u{ff9e}: 'ﾞ'

HalfwidthKatakanaSemiDashVoicedSoundMark

\u{ff9f}: 'ﾟ'

HalfwidthHangulFiller

\u{ffa0}: 'ﾠ'

HalfwidthHangulLetterKiyeok

\u{ffa1}: 'ﾡ'

HalfwidthHangulLetterSsangkiyeok

\u{ffa2}: 'ﾢ'

HalfwidthHangulLetterKiyeokDashSios

\u{ffa3}: 'ﾣ'

HalfwidthHangulLetterNieun

\u{ffa4}: 'ﾤ'

HalfwidthHangulLetterNieunDashCieuc

\u{ffa5}: 'ﾥ'

HalfwidthHangulLetterNieunDashHieuh

\u{ffa6}: 'ﾦ'

HalfwidthHangulLetterTikeut

\u{ffa7}: 'ﾧ'

HalfwidthHangulLetterSsangtikeut

\u{ffa8}: 'ﾨ'

HalfwidthHangulLetterRieul

\u{ffa9}: 'ﾩ'

HalfwidthHangulLetterRieulDashKiyeok

\u{ffaa}: 'ﾪ'

HalfwidthHangulLetterRieulDashMieum

\u{ffab}: 'ﾫ'

HalfwidthHangulLetterRieulDashPieup

\u{ffac}: 'ﾬ'

HalfwidthHangulLetterRieulDashSios

\u{ffad}: 'ﾭ'

HalfwidthHangulLetterRieulDashThieuth

\u{ffae}: 'ﾮ'

HalfwidthHangulLetterRieulDashPhieuph

\u{ffaf}: 'ﾯ'

HalfwidthHangulLetterRieulDashHieuh

\u{ffb0}: 'ﾰ'

HalfwidthHangulLetterMieum

\u{ffb1}: 'ﾱ'

HalfwidthHangulLetterPieup

\u{ffb2}: 'ﾲ'

HalfwidthHangulLetterSsangpieup

\u{ffb3}: 'ﾳ'

HalfwidthHangulLetterPieupDashSios

\u{ffb4}: 'ﾴ'

HalfwidthHangulLetterSios

\u{ffb5}: 'ﾵ'

HalfwidthHangulLetterSsangsios

\u{ffb6}: 'ﾶ'

HalfwidthHangulLetterIeung

\u{ffb7}: 'ﾷ'

HalfwidthHangulLetterCieuc

\u{ffb8}: 'ﾸ'

HalfwidthHangulLetterSsangcieuc

\u{ffb9}: 'ﾹ'

HalfwidthHangulLetterChieuch

\u{ffba}: 'ﾺ'

HalfwidthHangulLetterKhieukh

\u{ffbb}: 'ﾻ'

HalfwidthHangulLetterThieuth

\u{ffbc}: 'ﾼ'

HalfwidthHangulLetterPhieuph

\u{ffbd}: 'ﾽ'

HalfwidthHangulLetterHieuh

\u{ffbe}: 'ﾾ'

HalfwidthHangulLetterA

\u{ffc2}: 'ￂ'

HalfwidthHangulLetterAe

\u{ffc3}: 'ￃ'

HalfwidthHangulLetterYa

\u{ffc4}: 'ￄ'

HalfwidthHangulLetterYae

\u{ffc5}: 'ￅ'

HalfwidthHangulLetterEo

\u{ffc6}: 'ￆ'

HalfwidthHangulLetterE

\u{ffc7}: 'ￇ'

HalfwidthHangulLetterYeo

\u{ffca}: 'ￊ'

HalfwidthHangulLetterYe

\u{ffcb}: 'ￋ'

HalfwidthHangulLetterO

\u{ffcc}: 'ￌ'

HalfwidthHangulLetterWa

\u{ffcd}: 'ￍ'

HalfwidthHangulLetterWae

\u{ffce}: 'ￎ'

HalfwidthHangulLetterOe

\u{ffcf}: 'ￏ'

HalfwidthHangulLetterYo

\u{ffd2}: 'ￒ'

HalfwidthHangulLetterU

\u{ffd3}: 'ￓ'

HalfwidthHangulLetterWeo

\u{ffd4}: 'ￔ'

HalfwidthHangulLetterWe

\u{ffd5}: 'ￕ'

HalfwidthHangulLetterWi

\u{ffd6}: 'ￖ'

HalfwidthHangulLetterYu

\u{ffd7}: 'ￗ'

HalfwidthHangulLetterEu

\u{ffda}: 'ￚ'

HalfwidthHangulLetterYi

\u{ffdb}: 'ￛ'

HalfwidthHangulLetterI

\u{ffdc}: 'ￜ'

FullwidthCentSign

\u{ffe0}: '￠'

FullwidthPoundSign

\u{ffe1}: '￡'

FullwidthNotSign

\u{ffe2}: '￢'

FullwidthMacron

\u{ffe3}: '￣'

FullwidthBrokenBar

\u{ffe4}: '￤'

FullwidthYenSign

\u{ffe5}: '￥'

FullwidthWonSign

\u{ffe6}: '￦'

HalfwidthFormsLightVertical

\u{ffe8}: '￨'

HalfwidthLeftwardsArrow

\u{ffe9}: '￩'

HalfwidthUpwardsArrow

\u{ffea}: '￪'

HalfwidthRightwardsArrow

\u{ffeb}: '￫'

HalfwidthDownwardsArrow

\u{ffec}: '￬'

HalfwidthBlackSquare

\u{ffed}: '￭'

HalfwidthWhiteCircle

\u{ffee}: '￮'

Methods

impl HalfwidthandFullwidthForms

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case