Enum unicode_types::hangul_compatibility_jamo::HangulCompatibilityJamo

pub enum HangulCompatibilityJamo {
    HangulLetterKiyeok,
    HangulLetterSsangkiyeok,
    HangulLetterKiyeokDashSios,
    HangulLetterNieun,
    HangulLetterNieunDashCieuc,
    HangulLetterNieunDashHieuh,
    HangulLetterTikeut,
    HangulLetterSsangtikeut,
    HangulLetterRieul,
    HangulLetterRieulDashKiyeok,
    HangulLetterRieulDashMieum,
    HangulLetterRieulDashPieup,
    HangulLetterRieulDashSios,
    HangulLetterRieulDashThieuth,
    HangulLetterRieulDashPhieuph,
    HangulLetterRieulDashHieuh,
    HangulLetterMieum,
    HangulLetterPieup,
    HangulLetterSsangpieup,
    HangulLetterPieupDashSios,
    HangulLetterSios,
    HangulLetterSsangsios,
    HangulLetterIeung,
    HangulLetterCieuc,
    HangulLetterSsangcieuc,
    HangulLetterChieuch,
    HangulLetterKhieukh,
    HangulLetterThieuth,
    HangulLetterPhieuph,
    HangulLetterHieuh,
    HangulLetterA,
    HangulLetterAe,
    HangulLetterYa,
    HangulLetterYae,
    HangulLetterEo,
    HangulLetterE,
    HangulLetterYeo,
    HangulLetterYe,
    HangulLetterO,
    HangulLetterWa,
    HangulLetterWae,
    HangulLetterOe,
    HangulLetterYo,
    HangulLetterU,
    HangulLetterWeo,
    HangulLetterWe,
    HangulLetterWi,
    HangulLetterYu,
    HangulLetterEu,
    HangulLetterYi,
    HangulLetterI,
    HangulFiller,
    HangulLetterSsangnieun,
    HangulLetterNieunDashTikeut,
    HangulLetterNieunDashSios,
    HangulLetterNieunDashPansios,
    HangulLetterRieulDashKiyeokDashSios,
    HangulLetterRieulDashTikeut,
    HangulLetterRieulDashPieupDashSios,
    HangulLetterRieulDashPansios,
    HangulLetterRieulDashYeorinhieuh,
    HangulLetterMieumDashPieup,
    HangulLetterMieumDashSios,
    HangulLetterMieumDashPansios,
    HangulLetterKapyeounmieum,
    HangulLetterPieupDashKiyeok,
    HangulLetterPieupDashTikeut,
    HangulLetterPieupDashSiosDashKiyeok,
    HangulLetterPieupDashSiosDashTikeut,
    HangulLetterPieupDashCieuc,
    HangulLetterPieupDashThieuth,
    HangulLetterKapyeounpieup,
    HangulLetterKapyeounssangpieup,
    HangulLetterSiosDashKiyeok,
    HangulLetterSiosDashNieun,
    HangulLetterSiosDashTikeut,
    HangulLetterSiosDashPieup,
    HangulLetterSiosDashCieuc,
    HangulLetterPansios,
    HangulLetterSsangieung,
    HangulLetterYesieung,
    HangulLetterYesieungDashSios,
    HangulLetterYesieungDashPansios,
    HangulLetterKapyeounphieuph,
    HangulLetterSsanghieuh,
    HangulLetterYeorinhieuh,
    HangulLetterYoDashYa,
    HangulLetterYoDashYae,
    HangulLetterYoDashI,
    HangulLetterYuDashYeo,
    HangulLetterYuDashYe,
    HangulLetterYuDashI,
    HangulLetterAraea,
    HangulLetterAraeae,
}

An enum to represent all characters in the HangulCompatibilityJamo block.

Variants
HangulLetterKiyeok

\u{3131}: 'ㄱ'

HangulLetterSsangkiyeok

\u{3132}: 'ㄲ'

HangulLetterKiyeokDashSios

\u{3133}: 'ㄳ'

HangulLetterNieun

\u{3134}: 'ㄴ'

HangulLetterNieunDashCieuc

\u{3135}: 'ㄵ'

HangulLetterNieunDashHieuh

\u{3136}: 'ㄶ'

HangulLetterTikeut

\u{3137}: 'ㄷ'

HangulLetterSsangtikeut

\u{3138}: 'ㄸ'

HangulLetterRieul

\u{3139}: 'ㄹ'

HangulLetterRieulDashKiyeok

\u{313a}: 'ㄺ'

HangulLetterRieulDashMieum

\u{313b}: 'ㄻ'

HangulLetterRieulDashPieup

\u{313c}: 'ㄼ'

HangulLetterRieulDashSios

\u{313d}: 'ㄽ'

HangulLetterRieulDashThieuth

\u{313e}: 'ㄾ'

HangulLetterRieulDashPhieuph

\u{313f}: 'ㄿ'

HangulLetterRieulDashHieuh

\u{3140}: 'ㅀ'

HangulLetterMieum

\u{3141}: 'ㅁ'

HangulLetterPieup

\u{3142}: 'ㅂ'

HangulLetterSsangpieup

\u{3143}: 'ㅃ'

HangulLetterPieupDashSios

\u{3144}: 'ㅄ'

HangulLetterSios

\u{3145}: 'ㅅ'

HangulLetterSsangsios

\u{3146}: 'ㅆ'

HangulLetterIeung

\u{3147}: 'ㅇ'

HangulLetterCieuc

\u{3148}: 'ㅈ'

HangulLetterSsangcieuc

\u{3149}: 'ㅉ'

HangulLetterChieuch

\u{314a}: 'ㅊ'

HangulLetterKhieukh

\u{314b}: 'ㅋ'

HangulLetterThieuth

\u{314c}: 'ㅌ'

HangulLetterPhieuph

\u{314d}: 'ㅍ'

HangulLetterHieuh

\u{314e}: 'ㅎ'

HangulLetterA

\u{314f}: 'ㅏ'

HangulLetterAe

\u{3150}: 'ㅐ'

HangulLetterYa

\u{3151}: 'ㅑ'

HangulLetterYae

\u{3152}: 'ㅒ'

HangulLetterEo

\u{3153}: 'ㅓ'

HangulLetterE

\u{3154}: 'ㅔ'

HangulLetterYeo

\u{3155}: 'ㅕ'

HangulLetterYe

\u{3156}: 'ㅖ'

HangulLetterO

\u{3157}: 'ㅗ'

HangulLetterWa

\u{3158}: 'ㅘ'

HangulLetterWae

\u{3159}: 'ㅙ'

HangulLetterOe

\u{315a}: 'ㅚ'

HangulLetterYo

\u{315b}: 'ㅛ'

HangulLetterU

\u{315c}: 'ㅜ'

HangulLetterWeo

\u{315d}: 'ㅝ'

HangulLetterWe

\u{315e}: 'ㅞ'

HangulLetterWi

\u{315f}: 'ㅟ'

HangulLetterYu

\u{3160}: 'ㅠ'

HangulLetterEu

\u{3161}: 'ㅡ'

HangulLetterYi

\u{3162}: 'ㅢ'

HangulLetterI

\u{3163}: 'ㅣ'

HangulFiller

\u{3164}: 'ㅤ'

HangulLetterSsangnieun

\u{3165}: 'ㅥ'

HangulLetterNieunDashTikeut

\u{3166}: 'ㅦ'

HangulLetterNieunDashSios

\u{3167}: 'ㅧ'

HangulLetterNieunDashPansios

\u{3168}: 'ㅨ'

HangulLetterRieulDashKiyeokDashSios

\u{3169}: 'ㅩ'

HangulLetterRieulDashTikeut

\u{316a}: 'ㅪ'

HangulLetterRieulDashPieupDashSios

\u{316b}: 'ㅫ'

HangulLetterRieulDashPansios

\u{316c}: 'ㅬ'

HangulLetterRieulDashYeorinhieuh

\u{316d}: 'ㅭ'

HangulLetterMieumDashPieup

\u{316e}: 'ㅮ'

HangulLetterMieumDashSios

\u{316f}: 'ㅯ'

HangulLetterMieumDashPansios

\u{3170}: 'ㅰ'

HangulLetterKapyeounmieum

\u{3171}: 'ㅱ'

HangulLetterPieupDashKiyeok

\u{3172}: 'ㅲ'

HangulLetterPieupDashTikeut

\u{3173}: 'ㅳ'

HangulLetterPieupDashSiosDashKiyeok

\u{3174}: 'ㅴ'

HangulLetterPieupDashSiosDashTikeut

\u{3175}: 'ㅵ'

HangulLetterPieupDashCieuc

\u{3176}: 'ㅶ'

HangulLetterPieupDashThieuth

\u{3177}: 'ㅷ'

HangulLetterKapyeounpieup

\u{3178}: 'ㅸ'

HangulLetterKapyeounssangpieup

\u{3179}: 'ㅹ'

HangulLetterSiosDashKiyeok

\u{317a}: 'ㅺ'

HangulLetterSiosDashNieun

\u{317b}: 'ㅻ'

HangulLetterSiosDashTikeut

\u{317c}: 'ㅼ'

HangulLetterSiosDashPieup

\u{317d}: 'ㅽ'

HangulLetterSiosDashCieuc

\u{317e}: 'ㅾ'

HangulLetterPansios

\u{317f}: 'ㅿ'

HangulLetterSsangieung

\u{3180}: 'ㆀ'

HangulLetterYesieung

\u{3181}: 'ㆁ'

HangulLetterYesieungDashSios

\u{3182}: 'ㆂ'

HangulLetterYesieungDashPansios

\u{3183}: 'ㆃ'

HangulLetterKapyeounphieuph

\u{3184}: 'ㆄ'

HangulLetterSsanghieuh

\u{3185}: 'ㆅ'

HangulLetterYeorinhieuh

\u{3186}: 'ㆆ'

HangulLetterYoDashYa

\u{3187}: 'ㆇ'

HangulLetterYoDashYae

\u{3188}: 'ㆈ'

HangulLetterYoDashI

\u{3189}: 'ㆉ'

HangulLetterYuDashYeo

\u{318a}: 'ㆊ'

HangulLetterYuDashYe

\u{318b}: 'ㆋ'

HangulLetterYuDashI

\u{318c}: 'ㆌ'

HangulLetterAraea

\u{318d}: 'ㆍ'

HangulLetterAraeae

\u{318e}: 'ㆎ'

Methods

impl HangulCompatibilityJamo

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case