Enum unicode_types::hangul_jamo::HangulJamo

pub enum HangulJamo {
    HangulChoseongKiyeok,
    HangulChoseongSsangkiyeok,
    HangulChoseongNieun,
    HangulChoseongTikeut,
    HangulChoseongSsangtikeut,
    HangulChoseongRieul,
    HangulChoseongMieum,
    HangulChoseongPieup,
    HangulChoseongSsangpieup,
    HangulChoseongSios,
    HangulChoseongSsangsios,
    HangulChoseongIeung,
    HangulChoseongCieuc,
    HangulChoseongSsangcieuc,
    HangulChoseongChieuch,
    HangulChoseongKhieukh,
    HangulChoseongThieuth,
    HangulChoseongPhieuph,
    HangulChoseongHieuh,
    HangulChoseongNieunDashKiyeok,
    HangulChoseongSsangnieun,
    HangulChoseongNieunDashTikeut,
    HangulChoseongNieunDashPieup,
    HangulChoseongTikeutDashKiyeok,
    HangulChoseongRieulDashNieun,
    HangulChoseongSsangrieul,
    HangulChoseongRieulDashHieuh,
    HangulChoseongKapyeounrieul,
    HangulChoseongMieumDashPieup,
    HangulChoseongKapyeounmieum,
    HangulChoseongPieupDashKiyeok,
    HangulChoseongPieupDashNieun,
    HangulChoseongPieupDashTikeut,
    HangulChoseongPieupDashSios,
    HangulChoseongPieupDashSiosDashKiyeok,
    HangulChoseongPieupDashSiosDashTikeut,
    HangulChoseongPieupDashSiosDashPieup,
    HangulChoseongPieupDashSsangsios,
    HangulChoseongPieupDashSiosDashCieuc,
    HangulChoseongPieupDashCieuc,
    HangulChoseongPieupDashChieuch,
    HangulChoseongPieupDashThieuth,
    HangulChoseongPieupDashPhieuph,
    HangulChoseongKapyeounpieup,
    HangulChoseongKapyeounssangpieup,
    HangulChoseongSiosDashKiyeok,
    HangulChoseongSiosDashNieun,
    HangulChoseongSiosDashTikeut,
    HangulChoseongSiosDashRieul,
    HangulChoseongSiosDashMieum,
    HangulChoseongSiosDashPieup,
    HangulChoseongSiosDashPieupDashKiyeok,
    HangulChoseongSiosDashSsangsios,
    HangulChoseongSiosDashIeung,
    HangulChoseongSiosDashCieuc,
    HangulChoseongSiosDashChieuch,
    HangulChoseongSiosDashKhieukh,
    HangulChoseongSiosDashThieuth,
    HangulChoseongSiosDashPhieuph,
    HangulChoseongSiosDashHieuh,
    HangulChoseongChitueumsios,
    HangulChoseongChitueumssangsios,
    HangulChoseongCeongchieumsios,
    HangulChoseongCeongchieumssangsios,
    HangulChoseongPansios,
    HangulChoseongIeungDashKiyeok,
    HangulChoseongIeungDashTikeut,
    HangulChoseongIeungDashMieum,
    HangulChoseongIeungDashPieup,
    HangulChoseongIeungDashSios,
    HangulChoseongIeungDashPansios,
    HangulChoseongSsangieung,
    HangulChoseongIeungDashCieuc,
    HangulChoseongIeungDashChieuch,
    HangulChoseongIeungDashThieuth,
    HangulChoseongIeungDashPhieuph,
    HangulChoseongYesieung,
    HangulChoseongCieucDashIeung,
    HangulChoseongChitueumcieuc,
    HangulChoseongChitueumssangcieuc,
    HangulChoseongCeongchieumcieuc,
    HangulChoseongCeongchieumssangcieuc,
    HangulChoseongChieuchDashKhieukh,
    HangulChoseongChieuchDashHieuh,
    HangulChoseongChitueumchieuch,
    HangulChoseongCeongchieumchieuch,
    HangulChoseongPhieuphDashPieup,
    HangulChoseongKapyeounphieuph,
    HangulChoseongSsanghieuh,
    HangulChoseongYeorinhieuh,
    HangulChoseongKiyeokDashTikeut,
    HangulChoseongNieunDashSios,
    HangulChoseongNieunDashCieuc,
    HangulChoseongNieunDashHieuh,
    HangulChoseongTikeutDashRieul,
    HangulChoseongFiller,
    HangulJungseongFiller,
    HangulJungseongA,
    HangulJungseongAe,
    HangulJungseongYa,
    HangulJungseongYae,
    HangulJungseongEo,
    HangulJungseongE,
    HangulJungseongYeo,
    HangulJungseongYe,
    HangulJungseongO,
    HangulJungseongWa,
    HangulJungseongWae,
    HangulJungseongOe,
    HangulJungseongYo,
    HangulJungseongU,
    HangulJungseongWeo,
    HangulJungseongWe,
    HangulJungseongWi,
    HangulJungseongYu,
    HangulJungseongEu,
    HangulJungseongYi,
    HangulJungseongI,
    HangulJungseongADashO,
    HangulJungseongADashU,
    HangulJungseongYaDashO,
    HangulJungseongYaDashYo,
    HangulJungseongEoDashO,
    HangulJungseongEoDashU,
    HangulJungseongEoDashEu,
    HangulJungseongYeoDashO,
    HangulJungseongYeoDashU,
    HangulJungseongODashEo,
    HangulJungseongODashE,
    HangulJungseongODashYe,
    HangulJungseongODashO,
    HangulJungseongODashU,
    HangulJungseongYoDashYa,
    HangulJungseongYoDashYae,
    HangulJungseongYoDashYeo,
    HangulJungseongYoDashO,
    HangulJungseongYoDashI,
    HangulJungseongUDashA,
    HangulJungseongUDashAe,
    HangulJungseongUDashEoDashEu,
    HangulJungseongUDashYe,
    HangulJungseongUDashU,
    HangulJungseongYuDashA,
    HangulJungseongYuDashEo,
    HangulJungseongYuDashE,
    HangulJungseongYuDashYeo,
    HangulJungseongYuDashYe,
    HangulJungseongYuDashU,
    HangulJungseongYuDashI,
    HangulJungseongEuDashU,
    HangulJungseongEuDashEu,
    HangulJungseongYiDashU,
    HangulJungseongIDashA,
    HangulJungseongIDashYa,
    HangulJungseongIDashO,
    HangulJungseongIDashU,
    HangulJungseongIDashEu,
    HangulJungseongIDashAraea,
    HangulJungseongAraea,
    HangulJungseongAraeaDashEo,
    HangulJungseongAraeaDashU,
    HangulJungseongAraeaDashI,
    HangulJungseongSsangaraea,
    HangulJungseongADashEu,
    HangulJungseongYaDashU,
    HangulJungseongYeoDashYa,
    HangulJungseongODashYa,
    HangulJungseongODashYae,
    HangulJongseongKiyeok,
    HangulJongseongSsangkiyeok,
    HangulJongseongKiyeokDashSios,
    HangulJongseongNieun,
    HangulJongseongNieunDashCieuc,
    HangulJongseongNieunDashHieuh,
    HangulJongseongTikeut,
    HangulJongseongRieul,
    HangulJongseongRieulDashKiyeok,
    HangulJongseongRieulDashMieum,
    HangulJongseongRieulDashPieup,
    HangulJongseongRieulDashSios,
    HangulJongseongRieulDashThieuth,
    HangulJongseongRieulDashPhieuph,
    HangulJongseongRieulDashHieuh,
    HangulJongseongMieum,
    HangulJongseongPieup,
    HangulJongseongPieupDashSios,
    HangulJongseongSios,
    HangulJongseongSsangsios,
    HangulJongseongIeung,
    HangulJongseongCieuc,
    HangulJongseongChieuch,
    HangulJongseongKhieukh,
    HangulJongseongThieuth,
    HangulJongseongPhieuph,
    HangulJongseongHieuh,
    HangulJongseongKiyeokDashRieul,
    HangulJongseongKiyeokDashSiosDashKiyeok,
    HangulJongseongNieunDashKiyeok,
    HangulJongseongNieunDashTikeut,
    HangulJongseongNieunDashSios,
    HangulJongseongNieunDashPansios,
    HangulJongseongNieunDashThieuth,
    HangulJongseongTikeutDashKiyeok,
    HangulJongseongTikeutDashRieul,
    HangulJongseongRieulDashKiyeokDashSios,
    HangulJongseongRieulDashNieun,
    HangulJongseongRieulDashTikeut,
    HangulJongseongRieulDashTikeutDashHieuh,
    HangulJongseongSsangrieul,
    HangulJongseongRieulDashMieumDashKiyeok,
    HangulJongseongRieulDashMieumDashSios,
    HangulJongseongRieulDashPieupDashSios,
    HangulJongseongRieulDashPieupDashHieuh,
    HangulJongseongRieulDashKapyeounpieup,
    HangulJongseongRieulDashSsangsios,
    HangulJongseongRieulDashPansios,
    HangulJongseongRieulDashKhieukh,
    HangulJongseongRieulDashYeorinhieuh,
    HangulJongseongMieumDashKiyeok,
    HangulJongseongMieumDashRieul,
    HangulJongseongMieumDashPieup,
    HangulJongseongMieumDashSios,
    HangulJongseongMieumDashSsangsios,
    HangulJongseongMieumDashPansios,
    HangulJongseongMieumDashChieuch,
    HangulJongseongMieumDashHieuh,
    HangulJongseongKapyeounmieum,
    HangulJongseongPieupDashRieul,
    HangulJongseongPieupDashPhieuph,
    HangulJongseongPieupDashHieuh,
    HangulJongseongKapyeounpieup,
    HangulJongseongSiosDashKiyeok,
    HangulJongseongSiosDashTikeut,
    HangulJongseongSiosDashRieul,
    HangulJongseongSiosDashPieup,
    HangulJongseongPansios,
    HangulJongseongIeungDashKiyeok,
    HangulJongseongIeungDashSsangkiyeok,
    HangulJongseongSsangieung,
    HangulJongseongIeungDashKhieukh,
    HangulJongseongYesieung,
    HangulJongseongYesieungDashSios,
    HangulJongseongYesieungDashPansios,
    HangulJongseongPhieuphDashPieup,
    HangulJongseongKapyeounphieuph,
    HangulJongseongHieuhDashNieun,
    HangulJongseongHieuhDashRieul,
    HangulJongseongHieuhDashMieum,
    HangulJongseongHieuhDashPieup,
    HangulJongseongYeorinhieuh,
    HangulJongseongKiyeokDashNieun,
    HangulJongseongKiyeokDashPieup,
    HangulJongseongKiyeokDashChieuch,
    HangulJongseongKiyeokDashKhieukh,
    HangulJongseongKiyeokDashHieuh,
}

An enum to represent all characters in the HangulJamo block.

Variants
HangulChoseongKiyeok

\u{1100}: 'ᄀ'

HangulChoseongSsangkiyeok

\u{1101}: 'ᄁ'

HangulChoseongNieun

\u{1102}: 'ᄂ'

HangulChoseongTikeut

\u{1103}: 'ᄃ'

HangulChoseongSsangtikeut

\u{1104}: 'ᄄ'

HangulChoseongRieul

\u{1105}: 'ᄅ'

HangulChoseongMieum

\u{1106}: 'ᄆ'

HangulChoseongPieup

\u{1107}: 'ᄇ'

HangulChoseongSsangpieup

\u{1108}: 'ᄈ'

HangulChoseongSios

\u{1109}: 'ᄉ'

HangulChoseongSsangsios

\u{110a}: 'ᄊ'

HangulChoseongIeung

\u{110b}: 'ᄋ'

HangulChoseongCieuc

\u{110c}: 'ᄌ'

HangulChoseongSsangcieuc

\u{110d}: 'ᄍ'

HangulChoseongChieuch

\u{110e}: 'ᄎ'

HangulChoseongKhieukh

\u{110f}: 'ᄏ'

HangulChoseongThieuth

\u{1110}: 'ᄐ'

HangulChoseongPhieuph

\u{1111}: 'ᄑ'

HangulChoseongHieuh

\u{1112}: 'ᄒ'

HangulChoseongNieunDashKiyeok

\u{1113}: 'ᄓ'

HangulChoseongSsangnieun

\u{1114}: 'ᄔ'

HangulChoseongNieunDashTikeut

\u{1115}: 'ᄕ'

HangulChoseongNieunDashPieup

\u{1116}: 'ᄖ'

HangulChoseongTikeutDashKiyeok

\u{1117}: 'ᄗ'

HangulChoseongRieulDashNieun

\u{1118}: 'ᄘ'

HangulChoseongSsangrieul

\u{1119}: 'ᄙ'

HangulChoseongRieulDashHieuh

\u{111a}: 'ᄚ'

HangulChoseongKapyeounrieul

\u{111b}: 'ᄛ'

HangulChoseongMieumDashPieup

\u{111c}: 'ᄜ'

HangulChoseongKapyeounmieum

\u{111d}: 'ᄝ'

HangulChoseongPieupDashKiyeok

\u{111e}: 'ᄞ'

HangulChoseongPieupDashNieun

\u{111f}: 'ᄟ'

HangulChoseongPieupDashTikeut

\u{1120}: 'ᄠ'

HangulChoseongPieupDashSios

\u{1121}: 'ᄡ'

HangulChoseongPieupDashSiosDashKiyeok

\u{1122}: 'ᄢ'

HangulChoseongPieupDashSiosDashTikeut

\u{1123}: 'ᄣ'

HangulChoseongPieupDashSiosDashPieup

\u{1124}: 'ᄤ'

HangulChoseongPieupDashSsangsios

\u{1125}: 'ᄥ'

HangulChoseongPieupDashSiosDashCieuc

\u{1126}: 'ᄦ'

HangulChoseongPieupDashCieuc

\u{1127}: 'ᄧ'

HangulChoseongPieupDashChieuch

\u{1128}: 'ᄨ'

HangulChoseongPieupDashThieuth

\u{1129}: 'ᄩ'

HangulChoseongPieupDashPhieuph

\u{112a}: 'ᄪ'

HangulChoseongKapyeounpieup

\u{112b}: 'ᄫ'

HangulChoseongKapyeounssangpieup

\u{112c}: 'ᄬ'

HangulChoseongSiosDashKiyeok

\u{112d}: 'ᄭ'

HangulChoseongSiosDashNieun

\u{112e}: 'ᄮ'

HangulChoseongSiosDashTikeut

\u{112f}: 'ᄯ'

HangulChoseongSiosDashRieul

\u{1130}: 'ᄰ'

HangulChoseongSiosDashMieum

\u{1131}: 'ᄱ'

HangulChoseongSiosDashPieup

\u{1132}: 'ᄲ'

HangulChoseongSiosDashPieupDashKiyeok

\u{1133}: 'ᄳ'

HangulChoseongSiosDashSsangsios

\u{1134}: 'ᄴ'

HangulChoseongSiosDashIeung

\u{1135}: 'ᄵ'

HangulChoseongSiosDashCieuc

\u{1136}: 'ᄶ'

HangulChoseongSiosDashChieuch

\u{1137}: 'ᄷ'

HangulChoseongSiosDashKhieukh

\u{1138}: 'ᄸ'

HangulChoseongSiosDashThieuth

\u{1139}: 'ᄹ'

HangulChoseongSiosDashPhieuph

\u{113a}: 'ᄺ'

HangulChoseongSiosDashHieuh

\u{113b}: 'ᄻ'

HangulChoseongChitueumsios

\u{113c}: 'ᄼ'

HangulChoseongChitueumssangsios

\u{113d}: 'ᄽ'

HangulChoseongCeongchieumsios

\u{113e}: 'ᄾ'

HangulChoseongCeongchieumssangsios

\u{113f}: 'ᄿ'

HangulChoseongPansios

\u{1140}: 'ᅀ'

HangulChoseongIeungDashKiyeok

\u{1141}: 'ᅁ'

HangulChoseongIeungDashTikeut

\u{1142}: 'ᅂ'

HangulChoseongIeungDashMieum

\u{1143}: 'ᅃ'

HangulChoseongIeungDashPieup

\u{1144}: 'ᅄ'

HangulChoseongIeungDashSios

\u{1145}: 'ᅅ'

HangulChoseongIeungDashPansios

\u{1146}: 'ᅆ'

HangulChoseongSsangieung

\u{1147}: 'ᅇ'

HangulChoseongIeungDashCieuc

\u{1148}: 'ᅈ'

HangulChoseongIeungDashChieuch

\u{1149}: 'ᅉ'

HangulChoseongIeungDashThieuth

\u{114a}: 'ᅊ'

HangulChoseongIeungDashPhieuph

\u{114b}: 'ᅋ'

HangulChoseongYesieung

\u{114c}: 'ᅌ'

HangulChoseongCieucDashIeung

\u{114d}: 'ᅍ'

HangulChoseongChitueumcieuc

\u{114e}: 'ᅎ'

HangulChoseongChitueumssangcieuc

\u{114f}: 'ᅏ'

HangulChoseongCeongchieumcieuc

\u{1150}: 'ᅐ'

HangulChoseongCeongchieumssangcieuc

\u{1151}: 'ᅑ'

HangulChoseongChieuchDashKhieukh

\u{1152}: 'ᅒ'

HangulChoseongChieuchDashHieuh

\u{1153}: 'ᅓ'

HangulChoseongChitueumchieuch

\u{1154}: 'ᅔ'

HangulChoseongCeongchieumchieuch

\u{1155}: 'ᅕ'

HangulChoseongPhieuphDashPieup

\u{1156}: 'ᅖ'

HangulChoseongKapyeounphieuph

\u{1157}: 'ᅗ'

HangulChoseongSsanghieuh

\u{1158}: 'ᅘ'

HangulChoseongYeorinhieuh

\u{1159}: 'ᅙ'

HangulChoseongKiyeokDashTikeut

\u{115a}: 'ᅚ'

HangulChoseongNieunDashSios

\u{115b}: 'ᅛ'

HangulChoseongNieunDashCieuc

\u{115c}: 'ᅜ'

HangulChoseongNieunDashHieuh

\u{115d}: 'ᅝ'

HangulChoseongTikeutDashRieul

\u{115e}: 'ᅞ'

HangulChoseongFiller

\u{115f}: 'ᅟ'

HangulJungseongFiller

\u{1160}: 'ᅠ'

HangulJungseongA

\u{1161}: 'ᅡ'

HangulJungseongAe

\u{1162}: 'ᅢ'

HangulJungseongYa

\u{1163}: 'ᅣ'

HangulJungseongYae

\u{1164}: 'ᅤ'

HangulJungseongEo

\u{1165}: 'ᅥ'

HangulJungseongE

\u{1166}: 'ᅦ'

HangulJungseongYeo

\u{1167}: 'ᅧ'

HangulJungseongYe

\u{1168}: 'ᅨ'

HangulJungseongO

\u{1169}: 'ᅩ'

HangulJungseongWa

\u{116a}: 'ᅪ'

HangulJungseongWae

\u{116b}: 'ᅫ'

HangulJungseongOe

\u{116c}: 'ᅬ'

HangulJungseongYo

\u{116d}: 'ᅭ'

HangulJungseongU

\u{116e}: 'ᅮ'

HangulJungseongWeo

\u{116f}: 'ᅯ'

HangulJungseongWe

\u{1170}: 'ᅰ'

HangulJungseongWi

\u{1171}: 'ᅱ'

HangulJungseongYu

\u{1172}: 'ᅲ'

HangulJungseongEu

\u{1173}: 'ᅳ'

HangulJungseongYi

\u{1174}: 'ᅴ'

HangulJungseongI

\u{1175}: 'ᅵ'

HangulJungseongADashO

\u{1176}: 'ᅶ'

HangulJungseongADashU

\u{1177}: 'ᅷ'

HangulJungseongYaDashO

\u{1178}: 'ᅸ'

HangulJungseongYaDashYo

\u{1179}: 'ᅹ'

HangulJungseongEoDashO

\u{117a}: 'ᅺ'

HangulJungseongEoDashU

\u{117b}: 'ᅻ'

HangulJungseongEoDashEu

\u{117c}: 'ᅼ'

HangulJungseongYeoDashO

\u{117d}: 'ᅽ'

HangulJungseongYeoDashU

\u{117e}: 'ᅾ'

HangulJungseongODashEo

\u{117f}: 'ᅿ'

HangulJungseongODashE

\u{1180}: 'ᆀ'

HangulJungseongODashYe

\u{1181}: 'ᆁ'

HangulJungseongODashO

\u{1182}: 'ᆂ'

HangulJungseongODashU

\u{1183}: 'ᆃ'

HangulJungseongYoDashYa

\u{1184}: 'ᆄ'

HangulJungseongYoDashYae

\u{1185}: 'ᆅ'

HangulJungseongYoDashYeo

\u{1186}: 'ᆆ'

HangulJungseongYoDashO

\u{1187}: 'ᆇ'

HangulJungseongYoDashI

\u{1188}: 'ᆈ'

HangulJungseongUDashA

\u{1189}: 'ᆉ'

HangulJungseongUDashAe

\u{118a}: 'ᆊ'

HangulJungseongUDashEoDashEu

\u{118b}: 'ᆋ'

HangulJungseongUDashYe

\u{118c}: 'ᆌ'

HangulJungseongUDashU

\u{118d}: 'ᆍ'

HangulJungseongYuDashA

\u{118e}: 'ᆎ'

HangulJungseongYuDashEo

\u{118f}: 'ᆏ'

HangulJungseongYuDashE

\u{1190}: 'ᆐ'

HangulJungseongYuDashYeo

\u{1191}: 'ᆑ'

HangulJungseongYuDashYe

\u{1192}: 'ᆒ'

HangulJungseongYuDashU

\u{1193}: 'ᆓ'

HangulJungseongYuDashI

\u{1194}: 'ᆔ'

HangulJungseongEuDashU

\u{1195}: 'ᆕ'

HangulJungseongEuDashEu

\u{1196}: 'ᆖ'

HangulJungseongYiDashU

\u{1197}: 'ᆗ'

HangulJungseongIDashA

\u{1198}: 'ᆘ'

HangulJungseongIDashYa

\u{1199}: 'ᆙ'

HangulJungseongIDashO

\u{119a}: 'ᆚ'

HangulJungseongIDashU

\u{119b}: 'ᆛ'

HangulJungseongIDashEu

\u{119c}: 'ᆜ'

HangulJungseongIDashAraea

\u{119d}: 'ᆝ'

HangulJungseongAraea

\u{119e}: 'ᆞ'

HangulJungseongAraeaDashEo

\u{119f}: 'ᆟ'

HangulJungseongAraeaDashU

\u{11a0}: 'ᆠ'

HangulJungseongAraeaDashI

\u{11a1}: 'ᆡ'

HangulJungseongSsangaraea

\u{11a2}: 'ᆢ'

HangulJungseongADashEu

\u{11a3}: 'ᆣ'

HangulJungseongYaDashU

\u{11a4}: 'ᆤ'

HangulJungseongYeoDashYa

\u{11a5}: 'ᆥ'

HangulJungseongODashYa

\u{11a6}: 'ᆦ'

HangulJungseongODashYae

\u{11a7}: 'ᆧ'

HangulJongseongKiyeok

\u{11a8}: 'ᆨ'

HangulJongseongSsangkiyeok

\u{11a9}: 'ᆩ'

HangulJongseongKiyeokDashSios

\u{11aa}: 'ᆪ'

HangulJongseongNieun

\u{11ab}: 'ᆫ'

HangulJongseongNieunDashCieuc

\u{11ac}: 'ᆬ'

HangulJongseongNieunDashHieuh

\u{11ad}: 'ᆭ'

HangulJongseongTikeut

\u{11ae}: 'ᆮ'

HangulJongseongRieul

\u{11af}: 'ᆯ'

HangulJongseongRieulDashKiyeok

\u{11b0}: 'ᆰ'

HangulJongseongRieulDashMieum

\u{11b1}: 'ᆱ'

HangulJongseongRieulDashPieup

\u{11b2}: 'ᆲ'

HangulJongseongRieulDashSios

\u{11b3}: 'ᆳ'

HangulJongseongRieulDashThieuth

\u{11b4}: 'ᆴ'

HangulJongseongRieulDashPhieuph

\u{11b5}: 'ᆵ'

HangulJongseongRieulDashHieuh

\u{11b6}: 'ᆶ'

HangulJongseongMieum

\u{11b7}: 'ᆷ'

HangulJongseongPieup

\u{11b8}: 'ᆸ'

HangulJongseongPieupDashSios

\u{11b9}: 'ᆹ'

HangulJongseongSios

\u{11ba}: 'ᆺ'

HangulJongseongSsangsios

\u{11bb}: 'ᆻ'

HangulJongseongIeung

\u{11bc}: 'ᆼ'

HangulJongseongCieuc

\u{11bd}: 'ᆽ'

HangulJongseongChieuch

\u{11be}: 'ᆾ'

HangulJongseongKhieukh

\u{11bf}: 'ᆿ'

HangulJongseongThieuth

\u{11c0}: 'ᇀ'

HangulJongseongPhieuph

\u{11c1}: 'ᇁ'

HangulJongseongHieuh

\u{11c2}: 'ᇂ'

HangulJongseongKiyeokDashRieul

\u{11c3}: 'ᇃ'

HangulJongseongKiyeokDashSiosDashKiyeok

\u{11c4}: 'ᇄ'

HangulJongseongNieunDashKiyeok

\u{11c5}: 'ᇅ'

HangulJongseongNieunDashTikeut

\u{11c6}: 'ᇆ'

HangulJongseongNieunDashSios

\u{11c7}: 'ᇇ'

HangulJongseongNieunDashPansios

\u{11c8}: 'ᇈ'

HangulJongseongNieunDashThieuth

\u{11c9}: 'ᇉ'

HangulJongseongTikeutDashKiyeok

\u{11ca}: 'ᇊ'

HangulJongseongTikeutDashRieul

\u{11cb}: 'ᇋ'

HangulJongseongRieulDashKiyeokDashSios

\u{11cc}: 'ᇌ'

HangulJongseongRieulDashNieun

\u{11cd}: 'ᇍ'

HangulJongseongRieulDashTikeut

\u{11ce}: 'ᇎ'

HangulJongseongRieulDashTikeutDashHieuh

\u{11cf}: 'ᇏ'

HangulJongseongSsangrieul

\u{11d0}: 'ᇐ'

HangulJongseongRieulDashMieumDashKiyeok

\u{11d1}: 'ᇑ'

HangulJongseongRieulDashMieumDashSios

\u{11d2}: 'ᇒ'

HangulJongseongRieulDashPieupDashSios

\u{11d3}: 'ᇓ'

HangulJongseongRieulDashPieupDashHieuh

\u{11d4}: 'ᇔ'

HangulJongseongRieulDashKapyeounpieup

\u{11d5}: 'ᇕ'

HangulJongseongRieulDashSsangsios

\u{11d6}: 'ᇖ'

HangulJongseongRieulDashPansios

\u{11d7}: 'ᇗ'

HangulJongseongRieulDashKhieukh

\u{11d8}: 'ᇘ'

HangulJongseongRieulDashYeorinhieuh

\u{11d9}: 'ᇙ'

HangulJongseongMieumDashKiyeok

\u{11da}: 'ᇚ'

HangulJongseongMieumDashRieul

\u{11db}: 'ᇛ'

HangulJongseongMieumDashPieup

\u{11dc}: 'ᇜ'

HangulJongseongMieumDashSios

\u{11dd}: 'ᇝ'

HangulJongseongMieumDashSsangsios

\u{11de}: 'ᇞ'

HangulJongseongMieumDashPansios

\u{11df}: 'ᇟ'

HangulJongseongMieumDashChieuch

\u{11e0}: 'ᇠ'

HangulJongseongMieumDashHieuh

\u{11e1}: 'ᇡ'

HangulJongseongKapyeounmieum

\u{11e2}: 'ᇢ'

HangulJongseongPieupDashRieul

\u{11e3}: 'ᇣ'

HangulJongseongPieupDashPhieuph

\u{11e4}: 'ᇤ'

HangulJongseongPieupDashHieuh

\u{11e5}: 'ᇥ'

HangulJongseongKapyeounpieup

\u{11e6}: 'ᇦ'

HangulJongseongSiosDashKiyeok

\u{11e7}: 'ᇧ'

HangulJongseongSiosDashTikeut

\u{11e8}: 'ᇨ'

HangulJongseongSiosDashRieul

\u{11e9}: 'ᇩ'

HangulJongseongSiosDashPieup

\u{11ea}: 'ᇪ'

HangulJongseongPansios

\u{11eb}: 'ᇫ'

HangulJongseongIeungDashKiyeok

\u{11ec}: 'ᇬ'

HangulJongseongIeungDashSsangkiyeok

\u{11ed}: 'ᇭ'

HangulJongseongSsangieung

\u{11ee}: 'ᇮ'

HangulJongseongIeungDashKhieukh

\u{11ef}: 'ᇯ'

HangulJongseongYesieung

\u{11f0}: 'ᇰ'

HangulJongseongYesieungDashSios

\u{11f1}: 'ᇱ'

HangulJongseongYesieungDashPansios

\u{11f2}: 'ᇲ'

HangulJongseongPhieuphDashPieup

\u{11f3}: 'ᇳ'

HangulJongseongKapyeounphieuph

\u{11f4}: 'ᇴ'

HangulJongseongHieuhDashNieun

\u{11f5}: 'ᇵ'

HangulJongseongHieuhDashRieul

\u{11f6}: 'ᇶ'

HangulJongseongHieuhDashMieum

\u{11f7}: 'ᇷ'

HangulJongseongHieuhDashPieup

\u{11f8}: 'ᇸ'

HangulJongseongYeorinhieuh

\u{11f9}: 'ᇹ'

HangulJongseongKiyeokDashNieun

\u{11fa}: 'ᇺ'

HangulJongseongKiyeokDashPieup

\u{11fb}: 'ᇻ'

HangulJongseongKiyeokDashChieuch

\u{11fc}: 'ᇼ'

HangulJongseongKiyeokDashKhieukh

\u{11fd}: 'ᇽ'

HangulJongseongKiyeokDashHieuh

\u{11fe}: 'ᇾ'

Methods

impl HangulJamo

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case