Enum unicode_types::kana_supplement::KanaSupplement

pub enum KanaSupplement {
    KatakanaLetterArchaicE,
    HiraganaLetterArchaicYe,
    HentaiganaLetterADash1,
    HentaiganaLetterADash2,
    HentaiganaLetterADash3,
    HentaiganaLetterADashWo,
    HentaiganaLetterIDash1,
    HentaiganaLetterIDash2,
    HentaiganaLetterIDash3,
    HentaiganaLetterIDash4,
    HentaiganaLetterUDash1,
    HentaiganaLetterUDash2,
    HentaiganaLetterUDash3,
    HentaiganaLetterUDash4,
    HentaiganaLetterUDash5,
    HentaiganaLetterEDash2,
    HentaiganaLetterEDash3,
    HentaiganaLetterEDash4,
    HentaiganaLetterEDash5,
    HentaiganaLetterEDash6,
    HentaiganaLetterODash1,
    HentaiganaLetterODash2,
    HentaiganaLetterODash3,
    HentaiganaLetterKaDash1,
    HentaiganaLetterKaDash2,
    HentaiganaLetterKaDash3,
    HentaiganaLetterKaDash4,
    HentaiganaLetterKaDash5,
    HentaiganaLetterKaDash6,
    HentaiganaLetterKaDash7,
    HentaiganaLetterKaDash8,
    HentaiganaLetterKaDash9,
    HentaiganaLetterKaDash10,
    HentaiganaLetterKaDash11,
    HentaiganaLetterKaDashKe,
    HentaiganaLetterKiDash1,
    HentaiganaLetterKiDash2,
    HentaiganaLetterKiDash3,
    HentaiganaLetterKiDash4,
    HentaiganaLetterKiDash5,
    HentaiganaLetterKiDash6,
    HentaiganaLetterKiDash7,
    HentaiganaLetterKiDash8,
    HentaiganaLetterKuDash1,
    HentaiganaLetterKuDash2,
    HentaiganaLetterKuDash3,
    HentaiganaLetterKuDash4,
    HentaiganaLetterKuDash5,
    HentaiganaLetterKuDash6,
    HentaiganaLetterKuDash7,
    HentaiganaLetterKeDash1,
    HentaiganaLetterKeDash2,
    HentaiganaLetterKeDash3,
    HentaiganaLetterKeDash4,
    HentaiganaLetterKeDash5,
    HentaiganaLetterKeDash6,
    HentaiganaLetterKoDash1,
    HentaiganaLetterKoDash2,
    HentaiganaLetterKoDash3,
    HentaiganaLetterKoDashKi,
    HentaiganaLetterSaDash1,
    HentaiganaLetterSaDash2,
    HentaiganaLetterSaDash3,
    HentaiganaLetterSaDash4,
    HentaiganaLetterSaDash5,
    HentaiganaLetterSaDash6,
    HentaiganaLetterSaDash7,
    HentaiganaLetterSaDash8,
    HentaiganaLetterSiDash1,
    HentaiganaLetterSiDash2,
    HentaiganaLetterSiDash3,
    HentaiganaLetterSiDash4,
    HentaiganaLetterSiDash5,
    HentaiganaLetterSiDash6,
    HentaiganaLetterSuDash1,
    HentaiganaLetterSuDash2,
    HentaiganaLetterSuDash3,
    HentaiganaLetterSuDash4,
    HentaiganaLetterSuDash5,
    HentaiganaLetterSuDash6,
    HentaiganaLetterSuDash7,
    HentaiganaLetterSuDash8,
    HentaiganaLetterSeDash1,
    HentaiganaLetterSeDash2,
    HentaiganaLetterSeDash3,
    HentaiganaLetterSeDash4,
    HentaiganaLetterSeDash5,
    HentaiganaLetterSoDash1,
    HentaiganaLetterSoDash2,
    HentaiganaLetterSoDash3,
    HentaiganaLetterSoDash4,
    HentaiganaLetterSoDash5,
    HentaiganaLetterSoDash6,
    HentaiganaLetterSoDash7,
    HentaiganaLetterTaDash1,
    HentaiganaLetterTaDash2,
    HentaiganaLetterTaDash3,
    HentaiganaLetterTaDash4,
    HentaiganaLetterTiDash1,
    HentaiganaLetterTiDash2,
    HentaiganaLetterTiDash3,
    HentaiganaLetterTiDash4,
    HentaiganaLetterTiDash5,
    HentaiganaLetterTiDash6,
    HentaiganaLetterTiDash7,
    HentaiganaLetterTuDash1,
    HentaiganaLetterTuDash2,
    HentaiganaLetterTuDash3,
    HentaiganaLetterTuDash4,
    HentaiganaLetterTuDashTo,
    HentaiganaLetterTeDash1,
    HentaiganaLetterTeDash2,
    HentaiganaLetterTeDash3,
    HentaiganaLetterTeDash4,
    HentaiganaLetterTeDash5,
    HentaiganaLetterTeDash6,
    HentaiganaLetterTeDash7,
    HentaiganaLetterTeDash8,
    HentaiganaLetterTeDash9,
    HentaiganaLetterToDash1,
    HentaiganaLetterToDash2,
    HentaiganaLetterToDash3,
    HentaiganaLetterToDash4,
    HentaiganaLetterToDash5,
    HentaiganaLetterToDash6,
    HentaiganaLetterToDashRa,
    HentaiganaLetterNaDash1,
    HentaiganaLetterNaDash2,
    HentaiganaLetterNaDash3,
    HentaiganaLetterNaDash4,
    HentaiganaLetterNaDash5,
    HentaiganaLetterNaDash6,
    HentaiganaLetterNaDash7,
    HentaiganaLetterNaDash8,
    HentaiganaLetterNaDash9,
    HentaiganaLetterNiDash1,
    HentaiganaLetterNiDash2,
    HentaiganaLetterNiDash3,
    HentaiganaLetterNiDash4,
    HentaiganaLetterNiDash5,
    HentaiganaLetterNiDash6,
    HentaiganaLetterNiDash7,
    HentaiganaLetterNiDashTe,
    HentaiganaLetterNuDash1,
    HentaiganaLetterNuDash2,
    HentaiganaLetterNuDash3,
    HentaiganaLetterNeDash1,
    HentaiganaLetterNeDash2,
    HentaiganaLetterNeDash3,
    HentaiganaLetterNeDash4,
    HentaiganaLetterNeDash5,
    HentaiganaLetterNeDash6,
    HentaiganaLetterNeDashKo,
    HentaiganaLetterNoDash1,
    HentaiganaLetterNoDash2,
    HentaiganaLetterNoDash3,
    HentaiganaLetterNoDash4,
    HentaiganaLetterNoDash5,
    HentaiganaLetterHaDash1,
    HentaiganaLetterHaDash2,
    HentaiganaLetterHaDash3,
    HentaiganaLetterHaDash4,
    HentaiganaLetterHaDash5,
    HentaiganaLetterHaDash6,
    HentaiganaLetterHaDash7,
    HentaiganaLetterHaDash8,
    HentaiganaLetterHaDash9,
    HentaiganaLetterHaDash10,
    HentaiganaLetterHaDash11,
    HentaiganaLetterHiDash1,
    HentaiganaLetterHiDash2,
    HentaiganaLetterHiDash3,
    HentaiganaLetterHiDash4,
    HentaiganaLetterHiDash5,
    HentaiganaLetterHiDash6,
    HentaiganaLetterHiDash7,
    HentaiganaLetterHuDash1,
    HentaiganaLetterHuDash2,
    HentaiganaLetterHuDash3,
    HentaiganaLetterHeDash1,
    HentaiganaLetterHeDash2,
    HentaiganaLetterHeDash3,
    HentaiganaLetterHeDash4,
    HentaiganaLetterHeDash5,
    HentaiganaLetterHeDash6,
    HentaiganaLetterHeDash7,
    HentaiganaLetterHoDash1,
    HentaiganaLetterHoDash2,
    HentaiganaLetterHoDash3,
    HentaiganaLetterHoDash4,
    HentaiganaLetterHoDash5,
    HentaiganaLetterHoDash6,
    HentaiganaLetterHoDash7,
    HentaiganaLetterHoDash8,
    HentaiganaLetterMaDash1,
    HentaiganaLetterMaDash2,
    HentaiganaLetterMaDash3,
    HentaiganaLetterMaDash4,
    HentaiganaLetterMaDash5,
    HentaiganaLetterMaDash6,
    HentaiganaLetterMaDash7,
    HentaiganaLetterMiDash1,
    HentaiganaLetterMiDash2,
    HentaiganaLetterMiDash3,
    HentaiganaLetterMiDash4,
    HentaiganaLetterMiDash5,
    HentaiganaLetterMiDash6,
    HentaiganaLetterMiDash7,
    HentaiganaLetterMuDash1,
    HentaiganaLetterMuDash2,
    HentaiganaLetterMuDash3,
    HentaiganaLetterMuDash4,
    HentaiganaLetterMeDash1,
    HentaiganaLetterMeDash2,
    HentaiganaLetterMeDashMa,
    HentaiganaLetterMoDash1,
    HentaiganaLetterMoDash2,
    HentaiganaLetterMoDash3,
    HentaiganaLetterMoDash4,
    HentaiganaLetterMoDash5,
    HentaiganaLetterMoDash6,
    HentaiganaLetterYaDash1,
    HentaiganaLetterYaDash2,
    HentaiganaLetterYaDash3,
    HentaiganaLetterYaDash4,
    HentaiganaLetterYaDash5,
    HentaiganaLetterYaDashYo,
    HentaiganaLetterYuDash1,
    HentaiganaLetterYuDash2,
    HentaiganaLetterYuDash3,
    HentaiganaLetterYuDash4,
    HentaiganaLetterYoDash1,
    HentaiganaLetterYoDash2,
    HentaiganaLetterYoDash3,
    HentaiganaLetterYoDash4,
    HentaiganaLetterYoDash5,
    HentaiganaLetterYoDash6,
    HentaiganaLetterRaDash1,
    HentaiganaLetterRaDash2,
    HentaiganaLetterRaDash3,
    HentaiganaLetterRaDash4,
    HentaiganaLetterRiDash1,
    HentaiganaLetterRiDash2,
    HentaiganaLetterRiDash3,
    HentaiganaLetterRiDash4,
    HentaiganaLetterRiDash5,
    HentaiganaLetterRiDash6,
    HentaiganaLetterRiDash7,
    HentaiganaLetterRuDash1,
    HentaiganaLetterRuDash2,
    HentaiganaLetterRuDash3,
    HentaiganaLetterRuDash4,
    HentaiganaLetterRuDash5,
    HentaiganaLetterRuDash6,
    HentaiganaLetterReDash1,
}

An enum to represent all characters in the KanaSupplement block.

Variants
KatakanaLetterArchaicE

\u{1b000}: 'ğ›€€'

HiraganaLetterArchaicYe

\u{1b001}: 'ğ›€'

HentaiganaLetterADash1

\u{1b002}: 'ğ›€‚'

HentaiganaLetterADash2

\u{1b003}: 'ğ›€ƒ'

HentaiganaLetterADash3

\u{1b004}: 'ğ›€„'

HentaiganaLetterADashWo

\u{1b005}: 'ğ›€…'

HentaiganaLetterIDash1

\u{1b006}: 'ğ›€†'

HentaiganaLetterIDash2

\u{1b007}: 'ğ›€‡'

HentaiganaLetterIDash3

\u{1b008}: 'ğ›€ˆ'

HentaiganaLetterIDash4

\u{1b009}: 'ğ›€‰'

HentaiganaLetterUDash1

\u{1b00a}: 'ğ›€Š'

HentaiganaLetterUDash2

\u{1b00b}: 'ğ›€‹'

HentaiganaLetterUDash3

\u{1b00c}: 'ğ›€Œ'

HentaiganaLetterUDash4

\u{1b00d}: 'ğ›€'

HentaiganaLetterUDash5

\u{1b00e}: 'ğ›€'

HentaiganaLetterEDash2

\u{1b00f}: 'ğ›€'

HentaiganaLetterEDash3

\u{1b010}: 'ğ›€'

HentaiganaLetterEDash4

\u{1b011}: 'ğ›€‘'

HentaiganaLetterEDash5

\u{1b012}: 'ğ›€’'

HentaiganaLetterEDash6

\u{1b013}: 'ğ›€“'

HentaiganaLetterODash1

\u{1b014}: 'ğ›€”'

HentaiganaLetterODash2

\u{1b015}: 'ğ›€•'

HentaiganaLetterODash3

\u{1b016}: 'ğ›€–'

HentaiganaLetterKaDash1

\u{1b017}: 'ğ›€—'

HentaiganaLetterKaDash2

\u{1b018}: 'ğ›€˜'

HentaiganaLetterKaDash3

\u{1b019}: 'ğ›€™'

HentaiganaLetterKaDash4

\u{1b01a}: 'ğ›€š'

HentaiganaLetterKaDash5

\u{1b01b}: 'ğ›€›'

HentaiganaLetterKaDash6

\u{1b01c}: 'ğ›€œ'

HentaiganaLetterKaDash7

\u{1b01d}: 'ğ›€'

HentaiganaLetterKaDash8

\u{1b01e}: 'ğ›€'

HentaiganaLetterKaDash9

\u{1b01f}: 'ğ›€Ÿ'

HentaiganaLetterKaDash10

\u{1b020}: 'ğ›€ '

HentaiganaLetterKaDash11

\u{1b021}: 'ğ›€¡'

HentaiganaLetterKaDashKe

\u{1b022}: 'ğ›€¢'

HentaiganaLetterKiDash1

\u{1b023}: 'ğ›€£'

HentaiganaLetterKiDash2

\u{1b024}: 'ğ›€¤'

HentaiganaLetterKiDash3

\u{1b025}: 'ğ›€¥'

HentaiganaLetterKiDash4

\u{1b026}: 'ğ›€¦'

HentaiganaLetterKiDash5

\u{1b027}: 'ğ›€'

HentaiganaLetterKiDash6

\u{1b028}: 'ğ›€¨'

HentaiganaLetterKiDash7

\u{1b029}: 'ğ›€©'

HentaiganaLetterKiDash8

\u{1b02a}: 'ğ›€ª'

HentaiganaLetterKuDash1

\u{1b02b}: 'ğ›€«'

HentaiganaLetterKuDash2

\u{1b02c}: 'ğ›€¬'

HentaiganaLetterKuDash3

\u{1b02d}: 'ğ›€­'

HentaiganaLetterKuDash4

\u{1b02e}: 'ğ›€®'

HentaiganaLetterKuDash5

\u{1b02f}: 'ğ›€¯'

HentaiganaLetterKuDash6

\u{1b030}: 'ğ›€°'

HentaiganaLetterKuDash7

\u{1b031}: 'ğ›€±'

HentaiganaLetterKeDash1

\u{1b032}: 'ğ›€²'

HentaiganaLetterKeDash2

\u{1b033}: 'ğ›€³'

HentaiganaLetterKeDash3

\u{1b034}: 'ğ›€´'

HentaiganaLetterKeDash4

\u{1b035}: 'ğ›€µ'

HentaiganaLetterKeDash5

\u{1b036}: 'ğ›€¶'

HentaiganaLetterKeDash6

\u{1b037}: 'ğ›€·'

HentaiganaLetterKoDash1

\u{1b038}: 'ğ›€¸'

HentaiganaLetterKoDash2

\u{1b039}: 'ğ›€¹'

HentaiganaLetterKoDash3

\u{1b03a}: 'ğ›€º'

HentaiganaLetterKoDashKi

\u{1b03b}: 'ğ›€»'

HentaiganaLetterSaDash1

\u{1b03c}: 'ğ›€¼'

HentaiganaLetterSaDash2

\u{1b03d}: 'ğ›€½'

HentaiganaLetterSaDash3

\u{1b03e}: 'ğ›€¾'

HentaiganaLetterSaDash4

\u{1b03f}: 'ğ›€¿'

HentaiganaLetterSaDash5

\u{1b040}: 'ğ›€'

HentaiganaLetterSaDash6

\u{1b041}: 'ğ›'

HentaiganaLetterSaDash7

\u{1b042}: 'ğ›‚'

HentaiganaLetterSaDash8

\u{1b043}: 'ğ›ƒ'

HentaiganaLetterSiDash1

\u{1b044}: 'ğ›„'

HentaiganaLetterSiDash2

\u{1b045}: 'ğ›…'

HentaiganaLetterSiDash3

\u{1b046}: 'ğ›†'

HentaiganaLetterSiDash4

\u{1b047}: 'ğ›‡'

HentaiganaLetterSiDash5

\u{1b048}: 'ğ›ˆ'

HentaiganaLetterSiDash6

\u{1b049}: 'ğ›‰'

HentaiganaLetterSuDash1

\u{1b04a}: 'ğ›Š'

HentaiganaLetterSuDash2

\u{1b04b}: 'ğ›‹'

HentaiganaLetterSuDash3

\u{1b04c}: 'ğ›Œ'

HentaiganaLetterSuDash4

\u{1b04d}: 'ğ›'

HentaiganaLetterSuDash5

\u{1b04e}: 'ğ›'

HentaiganaLetterSuDash6

\u{1b04f}: 'ğ›'

HentaiganaLetterSuDash7

\u{1b050}: 'ğ›'

HentaiganaLetterSuDash8

\u{1b051}: 'ğ›‘'

HentaiganaLetterSeDash1

\u{1b052}: 'ğ›’'

HentaiganaLetterSeDash2

\u{1b053}: 'ğ›“'

HentaiganaLetterSeDash3

\u{1b054}: 'ğ›”'

HentaiganaLetterSeDash4

\u{1b055}: 'ğ›•'

HentaiganaLetterSeDash5

\u{1b056}: 'ğ›–'

HentaiganaLetterSoDash1

\u{1b057}: 'ğ›—'

HentaiganaLetterSoDash2

\u{1b058}: 'ğ›˜'

HentaiganaLetterSoDash3

\u{1b059}: 'ğ›™'

HentaiganaLetterSoDash4

\u{1b05a}: 'ğ›š'

HentaiganaLetterSoDash5

\u{1b05b}: 'ğ››'

HentaiganaLetterSoDash6

\u{1b05c}: 'ğ›œ'

HentaiganaLetterSoDash7

\u{1b05d}: 'ğ›'

HentaiganaLetterTaDash1

\u{1b05e}: 'ğ›'

HentaiganaLetterTaDash2

\u{1b05f}: 'ğ›Ÿ'

HentaiganaLetterTaDash3

\u{1b060}: 'ğ› '

HentaiganaLetterTaDash4

\u{1b061}: 'ğ›¡'

HentaiganaLetterTiDash1

\u{1b062}: 'ğ›¢'

HentaiganaLetterTiDash2

\u{1b063}: 'ğ›£'

HentaiganaLetterTiDash3

\u{1b064}: 'ğ›¤'

HentaiganaLetterTiDash4

\u{1b065}: 'ğ›¥'

HentaiganaLetterTiDash5

\u{1b066}: 'ğ›¦'

HentaiganaLetterTiDash6

\u{1b067}: 'ğ›'

HentaiganaLetterTiDash7

\u{1b068}: 'ğ›¨'

HentaiganaLetterTuDash1

\u{1b069}: 'ğ›©'

HentaiganaLetterTuDash2

\u{1b06a}: 'ğ›ª'

HentaiganaLetterTuDash3

\u{1b06b}: 'ğ›«'

HentaiganaLetterTuDash4

\u{1b06c}: 'ğ›¬'

HentaiganaLetterTuDashTo

\u{1b06d}: 'ğ›­'

HentaiganaLetterTeDash1

\u{1b06e}: 'ğ›®'

HentaiganaLetterTeDash2

\u{1b06f}: 'ğ›¯'

HentaiganaLetterTeDash3

\u{1b070}: 'ğ›°'

HentaiganaLetterTeDash4

\u{1b071}: 'ğ›±'

HentaiganaLetterTeDash5

\u{1b072}: 'ğ›²'

HentaiganaLetterTeDash6

\u{1b073}: 'ğ›³'

HentaiganaLetterTeDash7

\u{1b074}: 'ğ›´'

HentaiganaLetterTeDash8

\u{1b075}: 'ğ›µ'

HentaiganaLetterTeDash9

\u{1b076}: 'ğ›¶'

HentaiganaLetterToDash1

\u{1b077}: 'ğ›·'

HentaiganaLetterToDash2

\u{1b078}: 'ğ›¸'

HentaiganaLetterToDash3

\u{1b079}: 'ğ›¹'

HentaiganaLetterToDash4

\u{1b07a}: 'ğ›º'

HentaiganaLetterToDash5

\u{1b07b}: 'ğ›»'

HentaiganaLetterToDash6

\u{1b07c}: 'ğ›¼'

HentaiganaLetterToDashRa

\u{1b07d}: 'ğ›½'

HentaiganaLetterNaDash1

\u{1b07e}: 'ğ›¾'

HentaiganaLetterNaDash2

\u{1b07f}: 'ğ›¿'

HentaiganaLetterNaDash3

\u{1b080}: 'ğ›‚€'

HentaiganaLetterNaDash4

\u{1b081}: 'ğ›‚'

HentaiganaLetterNaDash5

\u{1b082}: 'ğ›‚‚'

HentaiganaLetterNaDash6

\u{1b083}: 'ğ›‚ƒ'

HentaiganaLetterNaDash7

\u{1b084}: 'ğ›‚„'

HentaiganaLetterNaDash8

\u{1b085}: 'ğ›‚…'

HentaiganaLetterNaDash9

\u{1b086}: 'ğ›‚†'

HentaiganaLetterNiDash1

\u{1b087}: 'ğ›‚‡'

HentaiganaLetterNiDash2

\u{1b088}: 'ğ›‚ˆ'

HentaiganaLetterNiDash3

\u{1b089}: 'ğ›‚‰'

HentaiganaLetterNiDash4

\u{1b08a}: 'ğ›‚Š'

HentaiganaLetterNiDash5

\u{1b08b}: 'ğ›‚‹'

HentaiganaLetterNiDash6

\u{1b08c}: 'ğ›‚Œ'

HentaiganaLetterNiDash7

\u{1b08d}: 'ğ›‚'

HentaiganaLetterNiDashTe

\u{1b08e}: 'ğ›‚'

HentaiganaLetterNuDash1

\u{1b08f}: 'ğ›‚'

HentaiganaLetterNuDash2

\u{1b090}: 'ğ›‚'

HentaiganaLetterNuDash3

\u{1b091}: 'ğ›‚‘'

HentaiganaLetterNeDash1

\u{1b092}: 'ğ›‚’'

HentaiganaLetterNeDash2

\u{1b093}: 'ğ›‚“'

HentaiganaLetterNeDash3

\u{1b094}: 'ğ›‚”'

HentaiganaLetterNeDash4

\u{1b095}: 'ğ›‚•'

HentaiganaLetterNeDash5

\u{1b096}: 'ğ›‚–'

HentaiganaLetterNeDash6

\u{1b097}: 'ğ›‚—'

HentaiganaLetterNeDashKo

\u{1b098}: 'ğ›‚˜'

HentaiganaLetterNoDash1

\u{1b099}: 'ğ›‚™'

HentaiganaLetterNoDash2

\u{1b09a}: 'ğ›‚š'

HentaiganaLetterNoDash3

\u{1b09b}: 'ğ›‚›'

HentaiganaLetterNoDash4

\u{1b09c}: 'ğ›‚œ'

HentaiganaLetterNoDash5

\u{1b09d}: 'ğ›‚'

HentaiganaLetterHaDash1

\u{1b09e}: 'ğ›‚'

HentaiganaLetterHaDash2

\u{1b09f}: 'ğ›‚Ÿ'

HentaiganaLetterHaDash3

\u{1b0a0}: 'ğ›‚ '

HentaiganaLetterHaDash4

\u{1b0a1}: 'ğ›‚¡'

HentaiganaLetterHaDash5

\u{1b0a2}: 'ğ›‚¢'

HentaiganaLetterHaDash6

\u{1b0a3}: 'ğ›‚£'

HentaiganaLetterHaDash7

\u{1b0a4}: 'ğ›‚¤'

HentaiganaLetterHaDash8

\u{1b0a5}: 'ğ›‚¥'

HentaiganaLetterHaDash9

\u{1b0a6}: 'ğ›‚¦'

HentaiganaLetterHaDash10

\u{1b0a7}: 'ğ›‚'

HentaiganaLetterHaDash11

\u{1b0a8}: 'ğ›‚¨'

HentaiganaLetterHiDash1

\u{1b0a9}: 'ğ›‚©'

HentaiganaLetterHiDash2

\u{1b0aa}: 'ğ›‚ª'

HentaiganaLetterHiDash3

\u{1b0ab}: 'ğ›‚«'

HentaiganaLetterHiDash4

\u{1b0ac}: 'ğ›‚¬'

HentaiganaLetterHiDash5

\u{1b0ad}: 'ğ›‚­'

HentaiganaLetterHiDash6

\u{1b0ae}: 'ğ›‚®'

HentaiganaLetterHiDash7

\u{1b0af}: 'ğ›‚¯'

HentaiganaLetterHuDash1

\u{1b0b0}: 'ğ›‚°'

HentaiganaLetterHuDash2

\u{1b0b1}: 'ğ›‚±'

HentaiganaLetterHuDash3

\u{1b0b2}: 'ğ›‚²'

HentaiganaLetterHeDash1

\u{1b0b3}: 'ğ›‚³'

HentaiganaLetterHeDash2

\u{1b0b4}: 'ğ›‚´'

HentaiganaLetterHeDash3

\u{1b0b5}: 'ğ›‚µ'

HentaiganaLetterHeDash4

\u{1b0b6}: 'ğ›‚¶'

HentaiganaLetterHeDash5

\u{1b0b7}: 'ğ›‚·'

HentaiganaLetterHeDash6

\u{1b0b8}: 'ğ›‚¸'

HentaiganaLetterHeDash7

\u{1b0b9}: 'ğ›‚¹'

HentaiganaLetterHoDash1

\u{1b0ba}: 'ğ›‚º'

HentaiganaLetterHoDash2

\u{1b0bb}: 'ğ›‚»'

HentaiganaLetterHoDash3

\u{1b0bc}: 'ğ›‚¼'

HentaiganaLetterHoDash4

\u{1b0bd}: 'ğ›‚½'

HentaiganaLetterHoDash5

\u{1b0be}: 'ğ›‚¾'

HentaiganaLetterHoDash6

\u{1b0bf}: 'ğ›‚¿'

HentaiganaLetterHoDash7

\u{1b0c0}: 'ğ›ƒ€'

HentaiganaLetterHoDash8

\u{1b0c1}: 'ğ›ƒ'

HentaiganaLetterMaDash1

\u{1b0c2}: 'ğ›ƒ‚'

HentaiganaLetterMaDash2

\u{1b0c3}: 'ğ›ƒƒ'

HentaiganaLetterMaDash3

\u{1b0c4}: 'ğ›ƒ„'

HentaiganaLetterMaDash4

\u{1b0c5}: 'ğ›ƒ…'

HentaiganaLetterMaDash5

\u{1b0c6}: 'ğ›ƒ†'

HentaiganaLetterMaDash6

\u{1b0c7}: 'ğ›ƒ‡'

HentaiganaLetterMaDash7

\u{1b0c8}: 'ğ›ƒˆ'

HentaiganaLetterMiDash1

\u{1b0c9}: 'ğ›ƒ‰'

HentaiganaLetterMiDash2

\u{1b0ca}: 'ğ›ƒŠ'

HentaiganaLetterMiDash3

\u{1b0cb}: 'ğ›ƒ‹'

HentaiganaLetterMiDash4

\u{1b0cc}: 'ğ›ƒŒ'

HentaiganaLetterMiDash5

\u{1b0cd}: 'ğ›ƒ'

HentaiganaLetterMiDash6

\u{1b0ce}: 'ğ›ƒ'

HentaiganaLetterMiDash7

\u{1b0cf}: 'ğ›ƒ'

HentaiganaLetterMuDash1

\u{1b0d0}: 'ğ›ƒ'

HentaiganaLetterMuDash2

\u{1b0d1}: 'ğ›ƒ‘'

HentaiganaLetterMuDash3

\u{1b0d2}: 'ğ›ƒ’'

HentaiganaLetterMuDash4

\u{1b0d3}: 'ğ›ƒ“'

HentaiganaLetterMeDash1

\u{1b0d4}: 'ğ›ƒ”'

HentaiganaLetterMeDash2

\u{1b0d5}: 'ğ›ƒ•'

HentaiganaLetterMeDashMa

\u{1b0d6}: 'ğ›ƒ–'

HentaiganaLetterMoDash1

\u{1b0d7}: 'ğ›ƒ—'

HentaiganaLetterMoDash2

\u{1b0d8}: 'ğ›ƒ˜'

HentaiganaLetterMoDash3

\u{1b0d9}: 'ğ›ƒ™'

HentaiganaLetterMoDash4

\u{1b0da}: 'ğ›ƒš'

HentaiganaLetterMoDash5

\u{1b0db}: 'ğ›ƒ›'

HentaiganaLetterMoDash6

\u{1b0dc}: 'ğ›ƒœ'

HentaiganaLetterYaDash1

\u{1b0dd}: 'ğ›ƒ'

HentaiganaLetterYaDash2

\u{1b0de}: 'ğ›ƒ'

HentaiganaLetterYaDash3

\u{1b0df}: 'ğ›ƒŸ'

HentaiganaLetterYaDash4

\u{1b0e0}: 'ğ›ƒ '

HentaiganaLetterYaDash5

\u{1b0e1}: 'ğ›ƒ¡'

HentaiganaLetterYaDashYo

\u{1b0e2}: 'ğ›ƒ¢'

HentaiganaLetterYuDash1

\u{1b0e3}: 'ğ›ƒ£'

HentaiganaLetterYuDash2

\u{1b0e4}: 'ğ›ƒ¤'

HentaiganaLetterYuDash3

\u{1b0e5}: 'ğ›ƒ¥'

HentaiganaLetterYuDash4

\u{1b0e6}: 'ğ›ƒ¦'

HentaiganaLetterYoDash1

\u{1b0e7}: 'ğ›ƒ'

HentaiganaLetterYoDash2

\u{1b0e8}: 'ğ›ƒ¨'

HentaiganaLetterYoDash3

\u{1b0e9}: 'ğ›ƒ©'

HentaiganaLetterYoDash4

\u{1b0ea}: 'ğ›ƒª'

HentaiganaLetterYoDash5

\u{1b0eb}: 'ğ›ƒ«'

HentaiganaLetterYoDash6

\u{1b0ec}: 'ğ›ƒ¬'

HentaiganaLetterRaDash1

\u{1b0ed}: 'ğ›ƒ­'

HentaiganaLetterRaDash2

\u{1b0ee}: 'ğ›ƒ®'

HentaiganaLetterRaDash3

\u{1b0ef}: 'ğ›ƒ¯'

HentaiganaLetterRaDash4

\u{1b0f0}: 'ğ›ƒ°'

HentaiganaLetterRiDash1

\u{1b0f1}: 'ğ›ƒ±'

HentaiganaLetterRiDash2

\u{1b0f2}: 'ğ›ƒ²'

HentaiganaLetterRiDash3

\u{1b0f3}: 'ğ›ƒ³'

HentaiganaLetterRiDash4

\u{1b0f4}: 'ğ›ƒ´'

HentaiganaLetterRiDash5

\u{1b0f5}: 'ğ›ƒµ'

HentaiganaLetterRiDash6

\u{1b0f6}: 'ğ›ƒ¶'

HentaiganaLetterRiDash7

\u{1b0f7}: 'ğ›ƒ·'

HentaiganaLetterRuDash1

\u{1b0f8}: 'ğ›ƒ¸'

HentaiganaLetterRuDash2

\u{1b0f9}: 'ğ›ƒ¹'

HentaiganaLetterRuDash3

\u{1b0fa}: 'ğ›ƒº'

HentaiganaLetterRuDash4

\u{1b0fb}: 'ğ›ƒ»'

HentaiganaLetterRuDash5

\u{1b0fc}: 'ğ›ƒ¼'

HentaiganaLetterRuDash6

\u{1b0fd}: 'ğ›ƒ½'

HentaiganaLetterReDash1

\u{1b0fe}: 'ğ›ƒ¾'

Methods

impl KanaSupplement

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case