Enum unicode_types::kannada::Kannada

pub enum Kannada {
    SignSpacingCandrabindu,
    SignCandrabindu,
    SignAnusvara,
    SignVisarga,
    SignSiddham,
    LetterA,
    LetterAa,
    LetterI,
    LetterIi,
    LetterU,
    LetterUu,
    LetterVocalicR,
    LetterVocalicL,
    LetterE,
    LetterEe,
    LetterAi,
    LetterO,
    LetterOo,
    LetterAu,
    LetterKa,
    LetterKha,
    LetterGa,
    LetterGha,
    LetterNga,
    LetterCa,
    LetterCha,
    LetterJa,
    LetterJha,
    LetterNya,
    LetterTta,
    LetterTtha,
    LetterDda,
    LetterDdha,
    LetterNna,
    LetterTa,
    LetterTha,
    LetterDa,
    LetterDha,
    LetterNa,
    LetterPa,
    LetterPha,
    LetterBa,
    LetterBha,
    LetterMa,
    LetterYa,
    LetterRa,
    LetterRra,
    LetterLa,
    LetterLla,
    LetterVa,
    LetterSha,
    LetterSsa,
    LetterSa,
    LetterHa,
    SignNukta,
    SignAvagraha,
    VowelSignAa,
    VowelSignI,
    VowelSignIi,
    VowelSignU,
    VowelSignUu,
    VowelSignVocalicR,
    VowelSignVocalicRr,
    VowelSignE,
    VowelSignEe,
    VowelSignAi,
    VowelSignO,
    VowelSignOo,
    VowelSignAu,
    SignVirama,
    LengthMark,
    AiLengthMark,
    LetterFa,
    LetterVocalicRr,
    LetterVocalicLl,
    VowelSignVocalicL,
    VowelSignVocalicLl,
    DigitZero,
    DigitOne,
    DigitTwo,
    DigitThree,
    DigitFour,
    DigitFive,
    DigitSix,
    DigitSeven,
    DigitEight,
    DigitNine,
    SignJihvamuliya,
    SignUpadhmaniya,
}

An enum to represent all characters in the Kannada block.

Variants
SignSpacingCandrabindu

\u{c80}: 'ಀ'

SignCandrabindu

\u{c81}: 'ಁ'

SignAnusvara

\u{c82}: 'ಂ'

SignVisarga

\u{c83}: 'ಃ'

SignSiddham

\u{c84}: '಄'

LetterA

\u{c85}: 'ಅ'

LetterAa

\u{c86}: 'ಆ'

LetterI

\u{c87}: 'ಇ'

LetterIi

\u{c88}: 'ಈ'

LetterU

\u{c89}: 'ಉ'

LetterUu

\u{c8a}: 'ಊ'

LetterVocalicR

\u{c8b}: 'ಋ'

LetterVocalicL

\u{c8c}: 'ಌ'

LetterE

\u{c8e}: 'ಎ'

LetterEe

\u{c8f}: 'ಏ'

LetterAi

\u{c90}: 'ಐ'

LetterO

\u{c92}: 'ಒ'

LetterOo

\u{c93}: 'ಓ'

LetterAu

\u{c94}: 'ಔ'

LetterKa

\u{c95}: 'ಕ'

LetterKha

\u{c96}: 'ಖ'

LetterGa

\u{c97}: 'ಗ'

LetterGha

\u{c98}: 'ಘ'

LetterNga

\u{c99}: 'ಙ'

LetterCa

\u{c9a}: 'ಚ'

LetterCha

\u{c9b}: 'ಛ'

LetterJa

\u{c9c}: 'ಜ'

LetterJha

\u{c9d}: 'ಝ'

LetterNya

\u{c9e}: 'ಞ'

LetterTta

\u{c9f}: 'ಟ'

LetterTtha

\u{ca0}: 'ಠ'

LetterDda

\u{ca1}: 'ಡ'

LetterDdha

\u{ca2}: 'ಢ'

LetterNna

\u{ca3}: 'ಣ'

LetterTa

\u{ca4}: 'ತ'

LetterTha

\u{ca5}: 'ಥ'

LetterDa

\u{ca6}: 'ದ'

LetterDha

\u{ca7}: 'ಧ'

LetterNa

\u{ca8}: 'ನ'

LetterPa

\u{caa}: 'ಪ'

LetterPha

\u{cab}: 'ಫ'

LetterBa

\u{cac}: 'ಬ'

LetterBha

\u{cad}: 'ಭ'

LetterMa

\u{cae}: 'ಮ'

LetterYa

\u{caf}: 'ಯ'

LetterRa

\u{cb0}: 'ರ'

LetterRra

\u{cb1}: 'ಱ'

LetterLa

\u{cb2}: 'ಲ'

LetterLla

\u{cb3}: 'ಳ'

LetterVa

\u{cb5}: 'ವ'

LetterSha

\u{cb6}: 'ಶ'

LetterSsa

\u{cb7}: 'ಷ'

LetterSa

\u{cb8}: 'ಸ'

LetterHa

\u{cb9}: 'ಹ'

SignNukta

\u{cbc}: '಼'

SignAvagraha

\u{cbd}: 'ಽ'

VowelSignAa

\u{cbe}: 'ಾ'

VowelSignI

\u{cbf}: 'ಿ'

VowelSignIi

\u{cc0}: 'ೀ'

VowelSignU

\u{cc1}: 'ು'

VowelSignUu

\u{cc2}: 'ೂ'

VowelSignVocalicR

\u{cc3}: 'ೃ'

VowelSignVocalicRr

\u{cc4}: 'ೄ'

VowelSignE

\u{cc6}: 'ೆ'

VowelSignEe

\u{cc7}: 'ೇ'

VowelSignAi

\u{cc8}: 'ೈ'

VowelSignO

\u{cca}: 'ೊ'

VowelSignOo

\u{ccb}: 'ೋ'

VowelSignAu

\u{ccc}: 'ೌ'

SignVirama

\u{ccd}: '್'

LengthMark

\u{cd5}: 'ೕ'

AiLengthMark

\u{cd6}: 'ೖ'

LetterFa

\u{cde}: 'ೞ'

LetterVocalicRr

\u{ce0}: 'ೠ'

LetterVocalicLl

\u{ce1}: 'ೡ'

VowelSignVocalicL

\u{ce2}: 'ೢ'

VowelSignVocalicLl

\u{ce3}: 'ೣ'

DigitZero

\u{ce6}: '೦'

DigitOne

\u{ce7}: '೧'

DigitTwo

\u{ce8}: '೨'

DigitThree

\u{ce9}: '೩'

DigitFour

\u{cea}: '೪'

DigitFive

\u{ceb}: '೫'

DigitSix

\u{cec}: '೬'

DigitSeven

\u{ced}: '೭'

DigitEight

\u{cee}: '೮'

DigitNine

\u{cef}: '೯'

SignJihvamuliya

\u{cf1}: 'ೱ'

SignUpadhmaniya

\u{cf2}: 'ೲ'

Methods

impl Kannada

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case