Enum unicode_types::latin_1_supplement::Latin1Supplement

pub enum Latin1Supplement {
    Control0080,
    Control0081,
    Control0082,
    Control0083,
    Control0084,
    Control0085,
    Control0086,
    Control0087,
    Control0088,
    Control0089,
    Control008a,
    Control008b,
    Control008c,
    Control008d,
    Control008e,
    Control008f,
    Control0090,
    Control0091,
    Control0092,
    Control0093,
    Control0094,
    Control0095,
    Control0096,
    Control0097,
    Control0098,
    Control0099,
    Control009a,
    Control009b,
    Control009c,
    Control009d,
    Control009e,
    Control009f,
    NoDashBreakSpace,
    InvertedExclamationMark,
    CentSign,
    PoundSign,
    CurrencySign,
    YenSign,
    BrokenBar,
    SectionSign,
    Diaeresis,
    CopyrightSign,
    FeminineOrdinalIndicator,
    LeftDashPointingDoubleAngleQuotationMark,
    NotSign,
    SoftHyphen,
    RegisteredSign,
    Macron,
    DegreeSign,
    PlusDashMinusSign,
    SuperscriptTwo,
    SuperscriptThree,
    AcuteAccent,
    MicroSign,
    PilcrowSign,
    MiddleDot,
    Cedilla,
    SuperscriptOne,
    MasculineOrdinalIndicator,
    RightDashPointingDoubleAngleQuotationMark,
    VulgarFractionOneQuarter,
    VulgarFractionOneHalf,
    VulgarFractionThreeQuarters,
    InvertedQuestionMark,
    LatinCapitalLetterAWithGrave,
    LatinCapitalLetterAWithAcute,
    LatinCapitalLetterAWithCircumflex,
    LatinCapitalLetterAWithTilde,
    LatinCapitalLetterAWithDiaeresis,
    LatinCapitalLetterAWithRingAbove,
    LatinCapitalLetterAe,
    LatinCapitalLetterCWithCedilla,
    LatinCapitalLetterEWithGrave,
    LatinCapitalLetterEWithAcute,
    LatinCapitalLetterEWithCircumflex,
    LatinCapitalLetterEWithDiaeresis,
    LatinCapitalLetterIWithGrave,
    LatinCapitalLetterIWithAcute,
    LatinCapitalLetterIWithCircumflex,
    LatinCapitalLetterIWithDiaeresis,
    LatinCapitalLetterEth,
    LatinCapitalLetterNWithTilde,
    LatinCapitalLetterOWithGrave,
    LatinCapitalLetterOWithAcute,
    LatinCapitalLetterOWithCircumflex,
    LatinCapitalLetterOWithTilde,
    LatinCapitalLetterOWithDiaeresis,
    MultiplicationSign,
    LatinCapitalLetterOWithStroke,
    LatinCapitalLetterUWithGrave,
    LatinCapitalLetterUWithAcute,
    LatinCapitalLetterUWithCircumflex,
    LatinCapitalLetterUWithDiaeresis,
    LatinCapitalLetterYWithAcute,
    LatinCapitalLetterThorn,
    LatinSmallLetterSharpS,
    LatinSmallLetterAWithGrave,
    LatinSmallLetterAWithAcute,
    LatinSmallLetterAWithCircumflex,
    LatinSmallLetterAWithTilde,
    LatinSmallLetterAWithDiaeresis,
    LatinSmallLetterAWithRingAbove,
    LatinSmallLetterAe,
    LatinSmallLetterCWithCedilla,
    LatinSmallLetterEWithGrave,
    LatinSmallLetterEWithAcute,
    LatinSmallLetterEWithCircumflex,
    LatinSmallLetterEWithDiaeresis,
    LatinSmallLetterIWithGrave,
    LatinSmallLetterIWithAcute,
    LatinSmallLetterIWithCircumflex,
    LatinSmallLetterIWithDiaeresis,
    LatinSmallLetterEth,
    LatinSmallLetterNWithTilde,
    LatinSmallLetterOWithGrave,
    LatinSmallLetterOWithAcute,
    LatinSmallLetterOWithCircumflex,
    LatinSmallLetterOWithTilde,
    LatinSmallLetterOWithDiaeresis,
    DivisionSign,
    LatinSmallLetterOWithStroke,
    LatinSmallLetterUWithGrave,
    LatinSmallLetterUWithAcute,
    LatinSmallLetterUWithCircumflex,
    LatinSmallLetterUWithDiaeresis,
    LatinSmallLetterYWithAcute,
    LatinSmallLetterThorn,
}

An enum to represent all characters in the Latin1Supplement block.

Variants
Control0080

\u{80}: ''

Control0081

\u{81}: ''

Control0082

\u{82}: ''

Control0083

\u{83}: ''

Control0084

\u{84}: ''

Control0085

\u{85}: ' '

Control0086

\u{86}: ''

Control0087

\u{87}: ''

Control0088

\u{88}: ''

Control0089

\u{89}: ''

Control008a

\u{8a}: ''

Control008b

\u{8b}: ''

Control008c

\u{8c}: ''

Control008d

\u{8d}: ''

Control008e

\u{8e}: ''

Control008f

\u{8f}: ''

Control0090

\u{90}: ''

Control0091

\u{91}: ''

Control0092

\u{92}: ''

Control0093

\u{93}: ''

Control0094

\u{94}: ''

Control0095

\u{95}: ''

Control0096

\u{96}: ''

Control0097

\u{97}: ''

Control0098

\u{98}: ''

Control0099

\u{99}: ''

Control009a

\u{9a}: ''

Control009b

\u{9b}: ''

Control009c

\u{9c}: ''

Control009d

\u{9d}: ''

Control009e

\u{9e}: ''

Control009f

\u{9f}: ''

NoDashBreakSpace

\u{a0}: ' '

InvertedExclamationMark

\u{a1}: '¡'

CentSign

\u{a2}: '¢'

PoundSign

\u{a3}: '£'

CurrencySign

\u{a4}: '¤'

YenSign

\u{a5}: '¥'

BrokenBar

\u{a6}: '¦'

SectionSign

\u{a7}: ''

Diaeresis

\u{a8}: '¨'

CopyrightSign

\u{a9}: '©'

FeminineOrdinalIndicator

\u{aa}: 'ª'

LeftDashPointingDoubleAngleQuotationMark

\u{ab}: '«'

NotSign

\u{ac}: '¬'

SoftHyphen

\u{ad}: '­'

RegisteredSign

\u{ae}: '®'

Macron

\u{af}: '¯'

DegreeSign

\u{b0}: '°'

PlusDashMinusSign

\u{b1}: '±'

SuperscriptTwo

\u{b2}: '²'

SuperscriptThree

\u{b3}: '³'

AcuteAccent

\u{b4}: '´'

MicroSign

\u{b5}: 'µ'

PilcrowSign

\u{b6}: '¶'

MiddleDot

\u{b7}: '·'

Cedilla

\u{b8}: '¸'

SuperscriptOne

\u{b9}: '¹'

MasculineOrdinalIndicator

\u{ba}: 'º'

RightDashPointingDoubleAngleQuotationMark

\u{bb}: '»'

VulgarFractionOneQuarter

\u{bc}: '¼'

VulgarFractionOneHalf

\u{bd}: '½'

VulgarFractionThreeQuarters

\u{be}: '¾'

InvertedQuestionMark

\u{bf}: '¿'

LatinCapitalLetterAWithGrave

\u{c0}: 'À'

LatinCapitalLetterAWithAcute

\u{c1}: 'Á'

LatinCapitalLetterAWithCircumflex

\u{c2}: 'Â'

LatinCapitalLetterAWithTilde

\u{c3}: 'Ã'

LatinCapitalLetterAWithDiaeresis

\u{c4}: 'Ä'

LatinCapitalLetterAWithRingAbove

\u{c5}: 'Å'

LatinCapitalLetterAe

\u{c6}: 'Æ'

LatinCapitalLetterCWithCedilla

\u{c7}: 'Ç'

LatinCapitalLetterEWithGrave

\u{c8}: 'È'

LatinCapitalLetterEWithAcute

\u{c9}: 'É'

LatinCapitalLetterEWithCircumflex

\u{ca}: 'Ê'

LatinCapitalLetterEWithDiaeresis

\u{cb}: 'Ë'

LatinCapitalLetterIWithGrave

\u{cc}: 'Ì'

LatinCapitalLetterIWithAcute

\u{cd}: 'Í'

LatinCapitalLetterIWithCircumflex

\u{ce}: 'Î'

LatinCapitalLetterIWithDiaeresis

\u{cf}: 'Ï'

LatinCapitalLetterEth

\u{d0}: 'Ð'

LatinCapitalLetterNWithTilde

\u{d1}: 'Ñ'

LatinCapitalLetterOWithGrave

\u{d2}: 'Ò'

LatinCapitalLetterOWithAcute

\u{d3}: 'Ó'

LatinCapitalLetterOWithCircumflex

\u{d4}: 'Ô'

LatinCapitalLetterOWithTilde

\u{d5}: 'Õ'

LatinCapitalLetterOWithDiaeresis

\u{d6}: 'Ö'

MultiplicationSign

\u{d7}: '×'

LatinCapitalLetterOWithStroke

\u{d8}: 'Ø'

LatinCapitalLetterUWithGrave

\u{d9}: 'Ù'

LatinCapitalLetterUWithAcute

\u{da}: 'Ú'

LatinCapitalLetterUWithCircumflex

\u{db}: 'Û'

LatinCapitalLetterUWithDiaeresis

\u{dc}: 'Ü'

LatinCapitalLetterYWithAcute

\u{dd}: 'Ý'

LatinCapitalLetterThorn

\u{de}: 'Þ'

LatinSmallLetterSharpS

\u{df}: 'ß'

LatinSmallLetterAWithGrave

\u{e0}: 'à'

LatinSmallLetterAWithAcute

\u{e1}: 'á'

LatinSmallLetterAWithCircumflex

\u{e2}: 'â'

LatinSmallLetterAWithTilde

\u{e3}: 'ã'

LatinSmallLetterAWithDiaeresis

\u{e4}: 'ä'

LatinSmallLetterAWithRingAbove

\u{e5}: 'å'

LatinSmallLetterAe

\u{e6}: 'æ'

LatinSmallLetterCWithCedilla

\u{e7}: 'ç'

LatinSmallLetterEWithGrave

\u{e8}: 'è'

LatinSmallLetterEWithAcute

\u{e9}: 'é'

LatinSmallLetterEWithCircumflex

\u{ea}: 'ê'

LatinSmallLetterEWithDiaeresis

\u{eb}: 'ë'

LatinSmallLetterIWithGrave

\u{ec}: 'ì'

LatinSmallLetterIWithAcute

\u{ed}: 'í'

LatinSmallLetterIWithCircumflex

\u{ee}: 'î'

LatinSmallLetterIWithDiaeresis

\u{ef}: 'ï'

LatinSmallLetterEth

\u{f0}: 'ð'

LatinSmallLetterNWithTilde

\u{f1}: 'ñ'

LatinSmallLetterOWithGrave

\u{f2}: 'ò'

LatinSmallLetterOWithAcute

\u{f3}: 'ó'

LatinSmallLetterOWithCircumflex

\u{f4}: 'ô'

LatinSmallLetterOWithTilde

\u{f5}: 'õ'

LatinSmallLetterOWithDiaeresis

\u{f6}: 'ö'

DivisionSign

\u{f7}: '÷'

LatinSmallLetterOWithStroke

\u{f8}: 'ø'

LatinSmallLetterUWithGrave

\u{f9}: 'ù'

LatinSmallLetterUWithAcute

\u{fa}: 'ú'

LatinSmallLetterUWithCircumflex

\u{fb}: 'û'

LatinSmallLetterUWithDiaeresis

\u{fc}: 'ü'

LatinSmallLetterYWithAcute

\u{fd}: 'ý'

LatinSmallLetterThorn

\u{fe}: 'þ'

Methods

impl Latin1Supplement

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case