Enum unicode_types::latin_extended_a::LatinExtendedA

pub enum LatinExtendedA {
    LatinCapitalLetterAWithMacron,
    LatinSmallLetterAWithMacron,
    LatinCapitalLetterAWithBreve,
    LatinSmallLetterAWithBreve,
    LatinCapitalLetterAWithOgonek,
    LatinSmallLetterAWithOgonek,
    LatinCapitalLetterCWithAcute,
    LatinSmallLetterCWithAcute,
    LatinCapitalLetterCWithCircumflex,
    LatinSmallLetterCWithCircumflex,
    LatinCapitalLetterCWithDotAbove,
    LatinSmallLetterCWithDotAbove,
    LatinCapitalLetterCWithCaron,
    LatinSmallLetterCWithCaron,
    LatinCapitalLetterDWithCaron,
    LatinSmallLetterDWithCaron,
    LatinCapitalLetterDWithStroke,
    LatinSmallLetterDWithStroke,
    LatinCapitalLetterEWithMacron,
    LatinSmallLetterEWithMacron,
    LatinCapitalLetterEWithBreve,
    LatinSmallLetterEWithBreve,
    LatinCapitalLetterEWithDotAbove,
    LatinSmallLetterEWithDotAbove,
    LatinCapitalLetterEWithOgonek,
    LatinSmallLetterEWithOgonek,
    LatinCapitalLetterEWithCaron,
    LatinSmallLetterEWithCaron,
    LatinCapitalLetterGWithCircumflex,
    LatinSmallLetterGWithCircumflex,
    LatinCapitalLetterGWithBreve,
    LatinSmallLetterGWithBreve,
    LatinCapitalLetterGWithDotAbove,
    LatinSmallLetterGWithDotAbove,
    LatinCapitalLetterGWithCedilla,
    LatinSmallLetterGWithCedilla,
    LatinCapitalLetterHWithCircumflex,
    LatinSmallLetterHWithCircumflex,
    LatinCapitalLetterHWithStroke,
    LatinSmallLetterHWithStroke,
    LatinCapitalLetterIWithTilde,
    LatinSmallLetterIWithTilde,
    LatinCapitalLetterIWithMacron,
    LatinSmallLetterIWithMacron,
    LatinCapitalLetterIWithBreve,
    LatinSmallLetterIWithBreve,
    LatinCapitalLetterIWithOgonek,
    LatinSmallLetterIWithOgonek,
    LatinCapitalLetterIWithDotAbove,
    LatinSmallLetterDotlessI,
    LatinCapitalLigatureIj,
    LatinSmallLigatureIj,
    LatinCapitalLetterJWithCircumflex,
    LatinSmallLetterJWithCircumflex,
    LatinCapitalLetterKWithCedilla,
    LatinSmallLetterKWithCedilla,
    LatinSmallLetterKra,
    LatinCapitalLetterLWithAcute,
    LatinSmallLetterLWithAcute,
    LatinCapitalLetterLWithCedilla,
    LatinSmallLetterLWithCedilla,
    LatinCapitalLetterLWithCaron,
    LatinSmallLetterLWithCaron,
    LatinCapitalLetterLWithMiddleDot,
    LatinSmallLetterLWithMiddleDot,
    LatinCapitalLetterLWithStroke,
    LatinSmallLetterLWithStroke,
    LatinCapitalLetterNWithAcute,
    LatinSmallLetterNWithAcute,
    LatinCapitalLetterNWithCedilla,
    LatinSmallLetterNWithCedilla,
    LatinCapitalLetterNWithCaron,
    LatinSmallLetterNWithCaron,
    LatinSmallLetterNPrecededByApostrophe,
    LatinCapitalLetterEng,
    LatinSmallLetterEng,
    LatinCapitalLetterOWithMacron,
    LatinSmallLetterOWithMacron,
    LatinCapitalLetterOWithBreve,
    LatinSmallLetterOWithBreve,
    LatinCapitalLetterOWithDoubleAcute,
    LatinSmallLetterOWithDoubleAcute,
    LatinCapitalLigatureOe,
    LatinSmallLigatureOe,
    LatinCapitalLetterRWithAcute,
    LatinSmallLetterRWithAcute,
    LatinCapitalLetterRWithCedilla,
    LatinSmallLetterRWithCedilla,
    LatinCapitalLetterRWithCaron,
    LatinSmallLetterRWithCaron,
    LatinCapitalLetterSWithAcute,
    LatinSmallLetterSWithAcute,
    LatinCapitalLetterSWithCircumflex,
    LatinSmallLetterSWithCircumflex,
    LatinCapitalLetterSWithCedilla,
    LatinSmallLetterSWithCedilla,
    LatinCapitalLetterSWithCaron,
    LatinSmallLetterSWithCaron,
    LatinCapitalLetterTWithCedilla,
    LatinSmallLetterTWithCedilla,
    LatinCapitalLetterTWithCaron,
    LatinSmallLetterTWithCaron,
    LatinCapitalLetterTWithStroke,
    LatinSmallLetterTWithStroke,
    LatinCapitalLetterUWithTilde,
    LatinSmallLetterUWithTilde,
    LatinCapitalLetterUWithMacron,
    LatinSmallLetterUWithMacron,
    LatinCapitalLetterUWithBreve,
    LatinSmallLetterUWithBreve,
    LatinCapitalLetterUWithRingAbove,
    LatinSmallLetterUWithRingAbove,
    LatinCapitalLetterUWithDoubleAcute,
    LatinSmallLetterUWithDoubleAcute,
    LatinCapitalLetterUWithOgonek,
    LatinSmallLetterUWithOgonek,
    LatinCapitalLetterWWithCircumflex,
    LatinSmallLetterWWithCircumflex,
    LatinCapitalLetterYWithCircumflex,
    LatinSmallLetterYWithCircumflex,
    LatinCapitalLetterYWithDiaeresis,
    LatinCapitalLetterZWithAcute,
    LatinSmallLetterZWithAcute,
    LatinCapitalLetterZWithDotAbove,
    LatinSmallLetterZWithDotAbove,
    LatinCapitalLetterZWithCaron,
    LatinSmallLetterZWithCaron,
}

An enum to represent all characters in the LatinExtendedA block.

Variants
LatinCapitalLetterAWithMacron

\u{100}: 'Ā'

LatinSmallLetterAWithMacron

\u{101}: 'ā'

LatinCapitalLetterAWithBreve

\u{102}: 'Ă'

LatinSmallLetterAWithBreve

\u{103}: 'ă'

LatinCapitalLetterAWithOgonek

\u{104}: 'Ą'

LatinSmallLetterAWithOgonek

\u{105}: 'ą'

LatinCapitalLetterCWithAcute

\u{106}: 'Ć'

LatinSmallLetterCWithAcute

\u{107}: 'ć'

LatinCapitalLetterCWithCircumflex

\u{108}: 'Ĉ'

LatinSmallLetterCWithCircumflex

\u{109}: 'ĉ'

LatinCapitalLetterCWithDotAbove

\u{10a}: 'Ċ'

LatinSmallLetterCWithDotAbove

\u{10b}: 'ċ'

LatinCapitalLetterCWithCaron

\u{10c}: 'Č'

LatinSmallLetterCWithCaron

\u{10d}: 'č'

LatinCapitalLetterDWithCaron

\u{10e}: 'Ď'

LatinSmallLetterDWithCaron

\u{10f}: 'ď'

LatinCapitalLetterDWithStroke

\u{110}: 'Đ'

LatinSmallLetterDWithStroke

\u{111}: 'đ'

LatinCapitalLetterEWithMacron

\u{112}: 'Ē'

LatinSmallLetterEWithMacron

\u{113}: 'ē'

LatinCapitalLetterEWithBreve

\u{114}: 'Ĕ'

LatinSmallLetterEWithBreve

\u{115}: 'ĕ'

LatinCapitalLetterEWithDotAbove

\u{116}: 'Ė'

LatinSmallLetterEWithDotAbove

\u{117}: 'ė'

LatinCapitalLetterEWithOgonek

\u{118}: 'Ę'

LatinSmallLetterEWithOgonek

\u{119}: 'ę'

LatinCapitalLetterEWithCaron

\u{11a}: 'Ě'

LatinSmallLetterEWithCaron

\u{11b}: 'ě'

LatinCapitalLetterGWithCircumflex

\u{11c}: 'Ĝ'

LatinSmallLetterGWithCircumflex

\u{11d}: 'ĝ'

LatinCapitalLetterGWithBreve

\u{11e}: 'Ğ'

LatinSmallLetterGWithBreve

\u{11f}: 'ğ'

LatinCapitalLetterGWithDotAbove

\u{120}: 'Ġ'

LatinSmallLetterGWithDotAbove

\u{121}: 'ġ'

LatinCapitalLetterGWithCedilla

\u{122}: 'Ģ'

LatinSmallLetterGWithCedilla

\u{123}: 'ģ'

LatinCapitalLetterHWithCircumflex

\u{124}: 'Ĥ'

LatinSmallLetterHWithCircumflex

\u{125}: 'ĥ'

LatinCapitalLetterHWithStroke

\u{126}: 'Ħ'

LatinSmallLetterHWithStroke

\u{127}: 'ħ'

LatinCapitalLetterIWithTilde

\u{128}: 'Ĩ'

LatinSmallLetterIWithTilde

\u{129}: 'ĩ'

LatinCapitalLetterIWithMacron

\u{12a}: 'Ī'

LatinSmallLetterIWithMacron

\u{12b}: 'ī'

LatinCapitalLetterIWithBreve

\u{12c}: 'Ĭ'

LatinSmallLetterIWithBreve

\u{12d}: 'ĭ'

LatinCapitalLetterIWithOgonek

\u{12e}: 'Į'

LatinSmallLetterIWithOgonek

\u{12f}: 'į'

LatinCapitalLetterIWithDotAbove

\u{130}: 'İ'

LatinSmallLetterDotlessI

\u{131}: 'ı'

LatinCapitalLigatureIj

\u{132}: 'Ĳ'

LatinSmallLigatureIj

\u{133}: 'ĳ'

LatinCapitalLetterJWithCircumflex

\u{134}: 'Ĵ'

LatinSmallLetterJWithCircumflex

\u{135}: 'ĵ'

LatinCapitalLetterKWithCedilla

\u{136}: 'Ķ'

LatinSmallLetterKWithCedilla

\u{137}: 'ķ'

LatinSmallLetterKra

\u{138}: 'ĸ'

LatinCapitalLetterLWithAcute

\u{139}: 'Ĺ'

LatinSmallLetterLWithAcute

\u{13a}: 'ĺ'

LatinCapitalLetterLWithCedilla

\u{13b}: 'Ļ'

LatinSmallLetterLWithCedilla

\u{13c}: 'ļ'

LatinCapitalLetterLWithCaron

\u{13d}: 'Ľ'

LatinSmallLetterLWithCaron

\u{13e}: 'ľ'

LatinCapitalLetterLWithMiddleDot

\u{13f}: 'Ŀ'

LatinSmallLetterLWithMiddleDot

\u{140}: 'ŀ'

LatinCapitalLetterLWithStroke

\u{141}: 'Ł'

LatinSmallLetterLWithStroke

\u{142}: 'ł'

LatinCapitalLetterNWithAcute

\u{143}: 'Ń'

LatinSmallLetterNWithAcute

\u{144}: 'ń'

LatinCapitalLetterNWithCedilla

\u{145}: 'Ņ'

LatinSmallLetterNWithCedilla

\u{146}: 'ņ'

LatinCapitalLetterNWithCaron

\u{147}: 'Ň'

LatinSmallLetterNWithCaron

\u{148}: 'ň'

LatinSmallLetterNPrecededByApostrophe

\u{149}: 'ŉ'

LatinCapitalLetterEng

\u{14a}: 'Ŋ'

LatinSmallLetterEng

\u{14b}: 'ŋ'

LatinCapitalLetterOWithMacron

\u{14c}: 'Ō'

LatinSmallLetterOWithMacron

\u{14d}: 'ō'

LatinCapitalLetterOWithBreve

\u{14e}: 'Ŏ'

LatinSmallLetterOWithBreve

\u{14f}: 'ŏ'

LatinCapitalLetterOWithDoubleAcute

\u{150}: 'Ő'

LatinSmallLetterOWithDoubleAcute

\u{151}: 'ő'

LatinCapitalLigatureOe

\u{152}: 'Œ'

LatinSmallLigatureOe

\u{153}: 'œ'

LatinCapitalLetterRWithAcute

\u{154}: 'Ŕ'

LatinSmallLetterRWithAcute

\u{155}: 'ŕ'

LatinCapitalLetterRWithCedilla

\u{156}: 'Ŗ'

LatinSmallLetterRWithCedilla

\u{157}: 'ŗ'

LatinCapitalLetterRWithCaron

\u{158}: 'Ř'

LatinSmallLetterRWithCaron

\u{159}: 'ř'

LatinCapitalLetterSWithAcute

\u{15a}: 'Ś'

LatinSmallLetterSWithAcute

\u{15b}: 'ś'

LatinCapitalLetterSWithCircumflex

\u{15c}: 'Ŝ'

LatinSmallLetterSWithCircumflex

\u{15d}: 'ŝ'

LatinCapitalLetterSWithCedilla

\u{15e}: 'Ş'

LatinSmallLetterSWithCedilla

\u{15f}: 'ş'

LatinCapitalLetterSWithCaron

\u{160}: 'Š'

LatinSmallLetterSWithCaron

\u{161}: 'š'

LatinCapitalLetterTWithCedilla

\u{162}: 'Ţ'

LatinSmallLetterTWithCedilla

\u{163}: 'ţ'

LatinCapitalLetterTWithCaron

\u{164}: 'Ť'

LatinSmallLetterTWithCaron

\u{165}: 'ť'

LatinCapitalLetterTWithStroke

\u{166}: 'Ŧ'

LatinSmallLetterTWithStroke

\u{167}: 'ŧ'

LatinCapitalLetterUWithTilde

\u{168}: 'Ũ'

LatinSmallLetterUWithTilde

\u{169}: 'ũ'

LatinCapitalLetterUWithMacron

\u{16a}: 'Ū'

LatinSmallLetterUWithMacron

\u{16b}: 'ū'

LatinCapitalLetterUWithBreve

\u{16c}: 'Ŭ'

LatinSmallLetterUWithBreve

\u{16d}: 'ŭ'

LatinCapitalLetterUWithRingAbove

\u{16e}: 'Ů'

LatinSmallLetterUWithRingAbove

\u{16f}: 'ů'

LatinCapitalLetterUWithDoubleAcute

\u{170}: 'Ű'

LatinSmallLetterUWithDoubleAcute

\u{171}: 'ű'

LatinCapitalLetterUWithOgonek

\u{172}: 'Ų'

LatinSmallLetterUWithOgonek

\u{173}: 'ų'

LatinCapitalLetterWWithCircumflex

\u{174}: 'Ŵ'

LatinSmallLetterWWithCircumflex

\u{175}: 'ŵ'

LatinCapitalLetterYWithCircumflex

\u{176}: 'Ŷ'

LatinSmallLetterYWithCircumflex

\u{177}: 'ŷ'

LatinCapitalLetterYWithDiaeresis

\u{178}: 'Ÿ'

LatinCapitalLetterZWithAcute

\u{179}: 'Ź'

LatinSmallLetterZWithAcute

\u{17a}: 'ź'

LatinCapitalLetterZWithDotAbove

\u{17b}: 'Ż'

LatinSmallLetterZWithDotAbove

\u{17c}: 'ż'

LatinCapitalLetterZWithCaron

\u{17d}: 'Ž'

LatinSmallLetterZWithCaron

\u{17e}: 'ž'

Methods

impl LatinExtendedA

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case