Enum unicode_types::latin_extended_additional::LatinExtendedAdditional

pub enum LatinExtendedAdditional {
    LatinCapitalLetterAWithRingBelow,
    LatinSmallLetterAWithRingBelow,
    LatinCapitalLetterBWithDotAbove,
    LatinSmallLetterBWithDotAbove,
    LatinCapitalLetterBWithDotBelow,
    LatinSmallLetterBWithDotBelow,
    LatinCapitalLetterBWithLineBelow,
    LatinSmallLetterBWithLineBelow,
    LatinCapitalLetterCWithCedillaAndAcute,
    LatinSmallLetterCWithCedillaAndAcute,
    LatinCapitalLetterDWithDotAbove,
    LatinSmallLetterDWithDotAbove,
    LatinCapitalLetterDWithDotBelow,
    LatinSmallLetterDWithDotBelow,
    LatinCapitalLetterDWithLineBelow,
    LatinSmallLetterDWithLineBelow,
    LatinCapitalLetterDWithCedilla,
    LatinSmallLetterDWithCedilla,
    LatinCapitalLetterDWithCircumflexBelow,
    LatinSmallLetterDWithCircumflexBelow,
    LatinCapitalLetterEWithMacronAndGrave,
    LatinSmallLetterEWithMacronAndGrave,
    LatinCapitalLetterEWithMacronAndAcute,
    LatinSmallLetterEWithMacronAndAcute,
    LatinCapitalLetterEWithCircumflexBelow,
    LatinSmallLetterEWithCircumflexBelow,
    LatinCapitalLetterEWithTildeBelow,
    LatinSmallLetterEWithTildeBelow,
    LatinCapitalLetterEWithCedillaAndBreve,
    LatinSmallLetterEWithCedillaAndBreve,
    LatinCapitalLetterFWithDotAbove,
    LatinSmallLetterFWithDotAbove,
    LatinCapitalLetterGWithMacron,
    LatinSmallLetterGWithMacron,
    LatinCapitalLetterHWithDotAbove,
    LatinSmallLetterHWithDotAbove,
    LatinCapitalLetterHWithDotBelow,
    LatinSmallLetterHWithDotBelow,
    LatinCapitalLetterHWithDiaeresis,
    LatinSmallLetterHWithDiaeresis,
    LatinCapitalLetterHWithCedilla,
    LatinSmallLetterHWithCedilla,
    LatinCapitalLetterHWithBreveBelow,
    LatinSmallLetterHWithBreveBelow,
    LatinCapitalLetterIWithTildeBelow,
    LatinSmallLetterIWithTildeBelow,
    LatinCapitalLetterIWithDiaeresisAndAcute,
    LatinSmallLetterIWithDiaeresisAndAcute,
    LatinCapitalLetterKWithAcute,
    LatinSmallLetterKWithAcute,
    LatinCapitalLetterKWithDotBelow,
    LatinSmallLetterKWithDotBelow,
    LatinCapitalLetterKWithLineBelow,
    LatinSmallLetterKWithLineBelow,
    LatinCapitalLetterLWithDotBelow,
    LatinSmallLetterLWithDotBelow,
    LatinCapitalLetterLWithDotBelowAndMacron,
    LatinSmallLetterLWithDotBelowAndMacron,
    LatinCapitalLetterLWithLineBelow,
    LatinSmallLetterLWithLineBelow,
    LatinCapitalLetterLWithCircumflexBelow,
    LatinSmallLetterLWithCircumflexBelow,
    LatinCapitalLetterMWithAcute,
    LatinSmallLetterMWithAcute,
    LatinCapitalLetterMWithDotAbove,
    LatinSmallLetterMWithDotAbove,
    LatinCapitalLetterMWithDotBelow,
    LatinSmallLetterMWithDotBelow,
    LatinCapitalLetterNWithDotAbove,
    LatinSmallLetterNWithDotAbove,
    LatinCapitalLetterNWithDotBelow,
    LatinSmallLetterNWithDotBelow,
    LatinCapitalLetterNWithLineBelow,
    LatinSmallLetterNWithLineBelow,
    LatinCapitalLetterNWithCircumflexBelow,
    LatinSmallLetterNWithCircumflexBelow,
    LatinCapitalLetterOWithTildeAndAcute,
    LatinSmallLetterOWithTildeAndAcute,
    LatinCapitalLetterOWithTildeAndDiaeresis,
    LatinSmallLetterOWithTildeAndDiaeresis,
    LatinCapitalLetterOWithMacronAndGrave,
    LatinSmallLetterOWithMacronAndGrave,
    LatinCapitalLetterOWithMacronAndAcute,
    LatinSmallLetterOWithMacronAndAcute,
    LatinCapitalLetterPWithAcute,
    LatinSmallLetterPWithAcute,
    LatinCapitalLetterPWithDotAbove,
    LatinSmallLetterPWithDotAbove,
    LatinCapitalLetterRWithDotAbove,
    LatinSmallLetterRWithDotAbove,
    LatinCapitalLetterRWithDotBelow,
    LatinSmallLetterRWithDotBelow,
    LatinCapitalLetterRWithDotBelowAndMacron,
    LatinSmallLetterRWithDotBelowAndMacron,
    LatinCapitalLetterRWithLineBelow,
    LatinSmallLetterRWithLineBelow,
    LatinCapitalLetterSWithDotAbove,
    LatinSmallLetterSWithDotAbove,
    LatinCapitalLetterSWithDotBelow,
    LatinSmallLetterSWithDotBelow,
    LatinCapitalLetterSWithAcuteAndDotAbove,
    LatinSmallLetterSWithAcuteAndDotAbove,
    LatinCapitalLetterSWithCaronAndDotAbove,
    LatinSmallLetterSWithCaronAndDotAbove,
    LatinCapitalLetterSWithDotBelowAndDotAbove,
    LatinSmallLetterSWithDotBelowAndDotAbove,
    LatinCapitalLetterTWithDotAbove,
    LatinSmallLetterTWithDotAbove,
    LatinCapitalLetterTWithDotBelow,
    LatinSmallLetterTWithDotBelow,
    LatinCapitalLetterTWithLineBelow,
    LatinSmallLetterTWithLineBelow,
    LatinCapitalLetterTWithCircumflexBelow,
    LatinSmallLetterTWithCircumflexBelow,
    LatinCapitalLetterUWithDiaeresisBelow,
    LatinSmallLetterUWithDiaeresisBelow,
    LatinCapitalLetterUWithTildeBelow,
    LatinSmallLetterUWithTildeBelow,
    LatinCapitalLetterUWithCircumflexBelow,
    LatinSmallLetterUWithCircumflexBelow,
    LatinCapitalLetterUWithTildeAndAcute,
    LatinSmallLetterUWithTildeAndAcute,
    LatinCapitalLetterUWithMacronAndDiaeresis,
    LatinSmallLetterUWithMacronAndDiaeresis,
    LatinCapitalLetterVWithTilde,
    LatinSmallLetterVWithTilde,
    LatinCapitalLetterVWithDotBelow,
    LatinSmallLetterVWithDotBelow,
    LatinCapitalLetterWWithGrave,
    LatinSmallLetterWWithGrave,
    LatinCapitalLetterWWithAcute,
    LatinSmallLetterWWithAcute,
    LatinCapitalLetterWWithDiaeresis,
    LatinSmallLetterWWithDiaeresis,
    LatinCapitalLetterWWithDotAbove,
    LatinSmallLetterWWithDotAbove,
    LatinCapitalLetterWWithDotBelow,
    LatinSmallLetterWWithDotBelow,
    LatinCapitalLetterXWithDotAbove,
    LatinSmallLetterXWithDotAbove,
    LatinCapitalLetterXWithDiaeresis,
    LatinSmallLetterXWithDiaeresis,
    LatinCapitalLetterYWithDotAbove,
    LatinSmallLetterYWithDotAbove,
    LatinCapitalLetterZWithCircumflex,
    LatinSmallLetterZWithCircumflex,
    LatinCapitalLetterZWithDotBelow,
    LatinSmallLetterZWithDotBelow,
    LatinCapitalLetterZWithLineBelow,
    LatinSmallLetterZWithLineBelow,
    LatinSmallLetterHWithLineBelow,
    LatinSmallLetterTWithDiaeresis,
    LatinSmallLetterWWithRingAbove,
    LatinSmallLetterYWithRingAbove,
    LatinSmallLetterAWithRightHalfRing,
    LatinSmallLetterLongSWithDotAbove,
    LatinSmallLetterLongSWithDiagonalStroke,
    LatinSmallLetterLongSWithHighStroke,
    LatinCapitalLetterSharpS,
    LatinSmallLetterDelta,
    LatinCapitalLetterAWithDotBelow,
    LatinSmallLetterAWithDotBelow,
    LatinCapitalLetterAWithHookAbove,
    LatinSmallLetterAWithHookAbove,
    LatinCapitalLetterAWithCircumflexAndAcute,
    LatinSmallLetterAWithCircumflexAndAcute,
    LatinCapitalLetterAWithCircumflexAndGrave,
    LatinSmallLetterAWithCircumflexAndGrave,
    LatinCapitalLetterAWithCircumflexAndHookAbove,
    LatinSmallLetterAWithCircumflexAndHookAbove,
    LatinCapitalLetterAWithCircumflexAndTilde,
    LatinSmallLetterAWithCircumflexAndTilde,
    LatinCapitalLetterAWithCircumflexAndDotBelow,
    LatinSmallLetterAWithCircumflexAndDotBelow,
    LatinCapitalLetterAWithBreveAndAcute,
    LatinSmallLetterAWithBreveAndAcute,
    LatinCapitalLetterAWithBreveAndGrave,
    LatinSmallLetterAWithBreveAndGrave,
    LatinCapitalLetterAWithBreveAndHookAbove,
    LatinSmallLetterAWithBreveAndHookAbove,
    LatinCapitalLetterAWithBreveAndTilde,
    LatinSmallLetterAWithBreveAndTilde,
    LatinCapitalLetterAWithBreveAndDotBelow,
    LatinSmallLetterAWithBreveAndDotBelow,
    LatinCapitalLetterEWithDotBelow,
    LatinSmallLetterEWithDotBelow,
    LatinCapitalLetterEWithHookAbove,
    LatinSmallLetterEWithHookAbove,
    LatinCapitalLetterEWithTilde,
    LatinSmallLetterEWithTilde,
    LatinCapitalLetterEWithCircumflexAndAcute,
    LatinSmallLetterEWithCircumflexAndAcute,
    LatinCapitalLetterEWithCircumflexAndGrave,
    LatinSmallLetterEWithCircumflexAndGrave,
    LatinCapitalLetterEWithCircumflexAndHookAbove,
    LatinSmallLetterEWithCircumflexAndHookAbove,
    LatinCapitalLetterEWithCircumflexAndTilde,
    LatinSmallLetterEWithCircumflexAndTilde,
    LatinCapitalLetterEWithCircumflexAndDotBelow,
    LatinSmallLetterEWithCircumflexAndDotBelow,
    LatinCapitalLetterIWithHookAbove,
    LatinSmallLetterIWithHookAbove,
    LatinCapitalLetterIWithDotBelow,
    LatinSmallLetterIWithDotBelow,
    LatinCapitalLetterOWithDotBelow,
    LatinSmallLetterOWithDotBelow,
    LatinCapitalLetterOWithHookAbove,
    LatinSmallLetterOWithHookAbove,
    LatinCapitalLetterOWithCircumflexAndAcute,
    LatinSmallLetterOWithCircumflexAndAcute,
    LatinCapitalLetterOWithCircumflexAndGrave,
    LatinSmallLetterOWithCircumflexAndGrave,
    LatinCapitalLetterOWithCircumflexAndHookAbove,
    LatinSmallLetterOWithCircumflexAndHookAbove,
    LatinCapitalLetterOWithCircumflexAndTilde,
    LatinSmallLetterOWithCircumflexAndTilde,
    LatinCapitalLetterOWithCircumflexAndDotBelow,
    LatinSmallLetterOWithCircumflexAndDotBelow,
    LatinCapitalLetterOWithHornAndAcute,
    LatinSmallLetterOWithHornAndAcute,
    LatinCapitalLetterOWithHornAndGrave,
    LatinSmallLetterOWithHornAndGrave,
    LatinCapitalLetterOWithHornAndHookAbove,
    LatinSmallLetterOWithHornAndHookAbove,
    LatinCapitalLetterOWithHornAndTilde,
    LatinSmallLetterOWithHornAndTilde,
    LatinCapitalLetterOWithHornAndDotBelow,
    LatinSmallLetterOWithHornAndDotBelow,
    LatinCapitalLetterUWithDotBelow,
    LatinSmallLetterUWithDotBelow,
    LatinCapitalLetterUWithHookAbove,
    LatinSmallLetterUWithHookAbove,
    LatinCapitalLetterUWithHornAndAcute,
    LatinSmallLetterUWithHornAndAcute,
    LatinCapitalLetterUWithHornAndGrave,
    LatinSmallLetterUWithHornAndGrave,
    LatinCapitalLetterUWithHornAndHookAbove,
    LatinSmallLetterUWithHornAndHookAbove,
    LatinCapitalLetterUWithHornAndTilde,
    LatinSmallLetterUWithHornAndTilde,
    LatinCapitalLetterUWithHornAndDotBelow,
    LatinSmallLetterUWithHornAndDotBelow,
    LatinCapitalLetterYWithGrave,
    LatinSmallLetterYWithGrave,
    LatinCapitalLetterYWithDotBelow,
    LatinSmallLetterYWithDotBelow,
    LatinCapitalLetterYWithHookAbove,
    LatinSmallLetterYWithHookAbove,
    LatinCapitalLetterYWithTilde,
    LatinSmallLetterYWithTilde,
    LatinCapitalLetterMiddleDashWelshLl,
    LatinSmallLetterMiddleDashWelshLl,
    LatinCapitalLetterMiddleDashWelshV,
    LatinSmallLetterMiddleDashWelshV,
    LatinCapitalLetterYWithLoop,
}

An enum to represent all characters in the LatinExtendedAdditional block.

Variants
LatinCapitalLetterAWithRingBelow

\u{1e00}: 'Ḁ'

LatinSmallLetterAWithRingBelow

\u{1e01}: 'ḁ'

LatinCapitalLetterBWithDotAbove

\u{1e02}: 'Ḃ'

LatinSmallLetterBWithDotAbove

\u{1e03}: 'ḃ'

LatinCapitalLetterBWithDotBelow

\u{1e04}: 'Ḅ'

LatinSmallLetterBWithDotBelow

\u{1e05}: 'ḅ'

LatinCapitalLetterBWithLineBelow

\u{1e06}: 'Ḇ'

LatinSmallLetterBWithLineBelow

\u{1e07}: 'ḇ'

LatinCapitalLetterCWithCedillaAndAcute

\u{1e08}: 'Ḉ'

LatinSmallLetterCWithCedillaAndAcute

\u{1e09}: 'ḉ'

LatinCapitalLetterDWithDotAbove

\u{1e0a}: 'Ḋ'

LatinSmallLetterDWithDotAbove

\u{1e0b}: 'ḋ'

LatinCapitalLetterDWithDotBelow

\u{1e0c}: 'Ḍ'

LatinSmallLetterDWithDotBelow

\u{1e0d}: 'ḍ'

LatinCapitalLetterDWithLineBelow

\u{1e0e}: 'Ḏ'

LatinSmallLetterDWithLineBelow

\u{1e0f}: 'ḏ'

LatinCapitalLetterDWithCedilla

\u{1e10}: 'Ḑ'

LatinSmallLetterDWithCedilla

\u{1e11}: 'ḑ'

LatinCapitalLetterDWithCircumflexBelow

\u{1e12}: 'Ḓ'

LatinSmallLetterDWithCircumflexBelow

\u{1e13}: 'ḓ'

LatinCapitalLetterEWithMacronAndGrave

\u{1e14}: 'Ḕ'

LatinSmallLetterEWithMacronAndGrave

\u{1e15}: 'ḕ'

LatinCapitalLetterEWithMacronAndAcute

\u{1e16}: 'Ḗ'

LatinSmallLetterEWithMacronAndAcute

\u{1e17}: 'ḗ'

LatinCapitalLetterEWithCircumflexBelow

\u{1e18}: 'Ḙ'

LatinSmallLetterEWithCircumflexBelow

\u{1e19}: 'ḙ'

LatinCapitalLetterEWithTildeBelow

\u{1e1a}: 'Ḛ'

LatinSmallLetterEWithTildeBelow

\u{1e1b}: 'ḛ'

LatinCapitalLetterEWithCedillaAndBreve

\u{1e1c}: 'Ḝ'

LatinSmallLetterEWithCedillaAndBreve

\u{1e1d}: 'ḝ'

LatinCapitalLetterFWithDotAbove

\u{1e1e}: 'Ḟ'

LatinSmallLetterFWithDotAbove

\u{1e1f}: 'ḟ'

LatinCapitalLetterGWithMacron

\u{1e20}: 'Ḡ'

LatinSmallLetterGWithMacron

\u{1e21}: 'ḡ'

LatinCapitalLetterHWithDotAbove

\u{1e22}: 'Ḣ'

LatinSmallLetterHWithDotAbove

\u{1e23}: 'ḣ'

LatinCapitalLetterHWithDotBelow

\u{1e24}: 'Ḥ'

LatinSmallLetterHWithDotBelow

\u{1e25}: 'ḥ'

LatinCapitalLetterHWithDiaeresis

\u{1e26}: 'Ḧ'

LatinSmallLetterHWithDiaeresis

\u{1e27}: 'ḧ'

LatinCapitalLetterHWithCedilla

\u{1e28}: 'Ḩ'

LatinSmallLetterHWithCedilla

\u{1e29}: 'ḩ'

LatinCapitalLetterHWithBreveBelow

\u{1e2a}: 'Ḫ'

LatinSmallLetterHWithBreveBelow

\u{1e2b}: 'ḫ'

LatinCapitalLetterIWithTildeBelow

\u{1e2c}: 'Ḭ'

LatinSmallLetterIWithTildeBelow

\u{1e2d}: 'ḭ'

LatinCapitalLetterIWithDiaeresisAndAcute

\u{1e2e}: 'Ḯ'

LatinSmallLetterIWithDiaeresisAndAcute

\u{1e2f}: 'ḯ'

LatinCapitalLetterKWithAcute

\u{1e30}: 'Ḱ'

LatinSmallLetterKWithAcute

\u{1e31}: 'ḱ'

LatinCapitalLetterKWithDotBelow

\u{1e32}: 'Ḳ'

LatinSmallLetterKWithDotBelow

\u{1e33}: 'ḳ'

LatinCapitalLetterKWithLineBelow

\u{1e34}: 'Ḵ'

LatinSmallLetterKWithLineBelow

\u{1e35}: 'ḵ'

LatinCapitalLetterLWithDotBelow

\u{1e36}: 'Ḷ'

LatinSmallLetterLWithDotBelow

\u{1e37}: 'ḷ'

LatinCapitalLetterLWithDotBelowAndMacron

\u{1e38}: 'Ḹ'

LatinSmallLetterLWithDotBelowAndMacron

\u{1e39}: 'ḹ'

LatinCapitalLetterLWithLineBelow

\u{1e3a}: 'Ḻ'

LatinSmallLetterLWithLineBelow

\u{1e3b}: 'ḻ'

LatinCapitalLetterLWithCircumflexBelow

\u{1e3c}: 'Ḽ'

LatinSmallLetterLWithCircumflexBelow

\u{1e3d}: 'ḽ'

LatinCapitalLetterMWithAcute

\u{1e3e}: 'Ḿ'

LatinSmallLetterMWithAcute

\u{1e3f}: 'ḿ'

LatinCapitalLetterMWithDotAbove

\u{1e40}: 'Ṁ'

LatinSmallLetterMWithDotAbove

\u{1e41}: 'ṁ'

LatinCapitalLetterMWithDotBelow

\u{1e42}: 'Ṃ'

LatinSmallLetterMWithDotBelow

\u{1e43}: 'ṃ'

LatinCapitalLetterNWithDotAbove

\u{1e44}: 'Ṅ'

LatinSmallLetterNWithDotAbove

\u{1e45}: 'ṅ'

LatinCapitalLetterNWithDotBelow

\u{1e46}: 'Ṇ'

LatinSmallLetterNWithDotBelow

\u{1e47}: 'ṇ'

LatinCapitalLetterNWithLineBelow

\u{1e48}: 'Ṉ'

LatinSmallLetterNWithLineBelow

\u{1e49}: 'ṉ'

LatinCapitalLetterNWithCircumflexBelow

\u{1e4a}: 'Ṋ'

LatinSmallLetterNWithCircumflexBelow

\u{1e4b}: 'ṋ'

LatinCapitalLetterOWithTildeAndAcute

\u{1e4c}: 'Ṍ'

LatinSmallLetterOWithTildeAndAcute

\u{1e4d}: 'ṍ'

LatinCapitalLetterOWithTildeAndDiaeresis

\u{1e4e}: 'Ṏ'

LatinSmallLetterOWithTildeAndDiaeresis

\u{1e4f}: 'ṏ'

LatinCapitalLetterOWithMacronAndGrave

\u{1e50}: 'Ṑ'

LatinSmallLetterOWithMacronAndGrave

\u{1e51}: 'ṑ'

LatinCapitalLetterOWithMacronAndAcute

\u{1e52}: 'Ṓ'

LatinSmallLetterOWithMacronAndAcute

\u{1e53}: 'ṓ'

LatinCapitalLetterPWithAcute

\u{1e54}: 'Ṕ'

LatinSmallLetterPWithAcute

\u{1e55}: 'ṕ'

LatinCapitalLetterPWithDotAbove

\u{1e56}: 'Ṗ'

LatinSmallLetterPWithDotAbove

\u{1e57}: 'ṗ'

LatinCapitalLetterRWithDotAbove

\u{1e58}: 'Ṙ'

LatinSmallLetterRWithDotAbove

\u{1e59}: 'ṙ'

LatinCapitalLetterRWithDotBelow

\u{1e5a}: 'Ṛ'

LatinSmallLetterRWithDotBelow

\u{1e5b}: 'ṛ'

LatinCapitalLetterRWithDotBelowAndMacron

\u{1e5c}: 'Ṝ'

LatinSmallLetterRWithDotBelowAndMacron

\u{1e5d}: 'ṝ'

LatinCapitalLetterRWithLineBelow

\u{1e5e}: 'Ṟ'

LatinSmallLetterRWithLineBelow

\u{1e5f}: 'ṟ'

LatinCapitalLetterSWithDotAbove

\u{1e60}: 'Ṡ'

LatinSmallLetterSWithDotAbove

\u{1e61}: 'ṡ'

LatinCapitalLetterSWithDotBelow

\u{1e62}: 'Ṣ'

LatinSmallLetterSWithDotBelow

\u{1e63}: 'ṣ'

LatinCapitalLetterSWithAcuteAndDotAbove

\u{1e64}: 'Ṥ'

LatinSmallLetterSWithAcuteAndDotAbove

\u{1e65}: 'ṥ'

LatinCapitalLetterSWithCaronAndDotAbove

\u{1e66}: 'Ṧ'

LatinSmallLetterSWithCaronAndDotAbove

\u{1e67}: 'ṧ'

LatinCapitalLetterSWithDotBelowAndDotAbove

\u{1e68}: 'Ṩ'

LatinSmallLetterSWithDotBelowAndDotAbove

\u{1e69}: 'ṩ'

LatinCapitalLetterTWithDotAbove

\u{1e6a}: 'Ṫ'

LatinSmallLetterTWithDotAbove

\u{1e6b}: 'ṫ'

LatinCapitalLetterTWithDotBelow

\u{1e6c}: 'Ṭ'

LatinSmallLetterTWithDotBelow

\u{1e6d}: 'ṭ'

LatinCapitalLetterTWithLineBelow

\u{1e6e}: 'Ṯ'

LatinSmallLetterTWithLineBelow

\u{1e6f}: 'ṯ'

LatinCapitalLetterTWithCircumflexBelow

\u{1e70}: 'Ṱ'

LatinSmallLetterTWithCircumflexBelow

\u{1e71}: 'ṱ'

LatinCapitalLetterUWithDiaeresisBelow

\u{1e72}: 'Ṳ'

LatinSmallLetterUWithDiaeresisBelow

\u{1e73}: 'ṳ'

LatinCapitalLetterUWithTildeBelow

\u{1e74}: 'Ṵ'

LatinSmallLetterUWithTildeBelow

\u{1e75}: 'ṵ'

LatinCapitalLetterUWithCircumflexBelow

\u{1e76}: 'Ṷ'

LatinSmallLetterUWithCircumflexBelow

\u{1e77}: 'ṷ'

LatinCapitalLetterUWithTildeAndAcute

\u{1e78}: 'Ṹ'

LatinSmallLetterUWithTildeAndAcute

\u{1e79}: 'ṹ'

LatinCapitalLetterUWithMacronAndDiaeresis

\u{1e7a}: 'Ṻ'

LatinSmallLetterUWithMacronAndDiaeresis

\u{1e7b}: 'ṻ'

LatinCapitalLetterVWithTilde

\u{1e7c}: 'Ṽ'

LatinSmallLetterVWithTilde

\u{1e7d}: 'ṽ'

LatinCapitalLetterVWithDotBelow

\u{1e7e}: 'Ṿ'

LatinSmallLetterVWithDotBelow

\u{1e7f}: 'ṿ'

LatinCapitalLetterWWithGrave

\u{1e80}: 'Ẁ'

LatinSmallLetterWWithGrave

\u{1e81}: 'ẁ'

LatinCapitalLetterWWithAcute

\u{1e82}: 'Ẃ'

LatinSmallLetterWWithAcute

\u{1e83}: 'ẃ'

LatinCapitalLetterWWithDiaeresis

\u{1e84}: 'Ẅ'

LatinSmallLetterWWithDiaeresis

\u{1e85}: 'ẅ'

LatinCapitalLetterWWithDotAbove

\u{1e86}: 'Ẇ'

LatinSmallLetterWWithDotAbove

\u{1e87}: 'ẇ'

LatinCapitalLetterWWithDotBelow

\u{1e88}: 'Ẉ'

LatinSmallLetterWWithDotBelow

\u{1e89}: 'ẉ'

LatinCapitalLetterXWithDotAbove

\u{1e8a}: 'Ẋ'

LatinSmallLetterXWithDotAbove

\u{1e8b}: 'ẋ'

LatinCapitalLetterXWithDiaeresis

\u{1e8c}: 'Ẍ'

LatinSmallLetterXWithDiaeresis

\u{1e8d}: 'ẍ'

LatinCapitalLetterYWithDotAbove

\u{1e8e}: 'Ẏ'

LatinSmallLetterYWithDotAbove

\u{1e8f}: 'ẏ'

LatinCapitalLetterZWithCircumflex

\u{1e90}: 'Ẑ'

LatinSmallLetterZWithCircumflex

\u{1e91}: 'ẑ'

LatinCapitalLetterZWithDotBelow

\u{1e92}: 'Ẓ'

LatinSmallLetterZWithDotBelow

\u{1e93}: 'ẓ'

LatinCapitalLetterZWithLineBelow

\u{1e94}: 'Ẕ'

LatinSmallLetterZWithLineBelow

\u{1e95}: 'ẕ'

LatinSmallLetterHWithLineBelow

\u{1e96}: 'ẖ'

LatinSmallLetterTWithDiaeresis

\u{1e97}: 'ẗ'

LatinSmallLetterWWithRingAbove

\u{1e98}: 'ẘ'

LatinSmallLetterYWithRingAbove

\u{1e99}: 'ẙ'

LatinSmallLetterAWithRightHalfRing

\u{1e9a}: 'ẚ'

LatinSmallLetterLongSWithDotAbove

\u{1e9b}: 'ẛ'

LatinSmallLetterLongSWithDiagonalStroke

\u{1e9c}: 'ẜ'

LatinSmallLetterLongSWithHighStroke

\u{1e9d}: 'ẝ'

LatinCapitalLetterSharpS

\u{1e9e}: 'ẞ'

LatinSmallLetterDelta

\u{1e9f}: 'ẟ'

LatinCapitalLetterAWithDotBelow

\u{1ea0}: 'Ạ'

LatinSmallLetterAWithDotBelow

\u{1ea1}: 'ạ'

LatinCapitalLetterAWithHookAbove

\u{1ea2}: 'Ả'

LatinSmallLetterAWithHookAbove

\u{1ea3}: 'ả'

LatinCapitalLetterAWithCircumflexAndAcute

\u{1ea4}: 'Ấ'

LatinSmallLetterAWithCircumflexAndAcute

\u{1ea5}: 'ấ'

LatinCapitalLetterAWithCircumflexAndGrave

\u{1ea6}: 'Ầ'

LatinSmallLetterAWithCircumflexAndGrave

\u{1ea7}: 'ầ'

LatinCapitalLetterAWithCircumflexAndHookAbove

\u{1ea8}: 'Ẩ'

LatinSmallLetterAWithCircumflexAndHookAbove

\u{1ea9}: 'ẩ'

LatinCapitalLetterAWithCircumflexAndTilde

\u{1eaa}: 'Ẫ'

LatinSmallLetterAWithCircumflexAndTilde

\u{1eab}: 'ẫ'

LatinCapitalLetterAWithCircumflexAndDotBelow

\u{1eac}: 'Ậ'

LatinSmallLetterAWithCircumflexAndDotBelow

\u{1ead}: 'ậ'

LatinCapitalLetterAWithBreveAndAcute

\u{1eae}: 'Ắ'

LatinSmallLetterAWithBreveAndAcute

\u{1eaf}: 'ắ'

LatinCapitalLetterAWithBreveAndGrave

\u{1eb0}: 'Ằ'

LatinSmallLetterAWithBreveAndGrave

\u{1eb1}: 'ằ'

LatinCapitalLetterAWithBreveAndHookAbove

\u{1eb2}: 'Ẳ'

LatinSmallLetterAWithBreveAndHookAbove

\u{1eb3}: 'ẳ'

LatinCapitalLetterAWithBreveAndTilde

\u{1eb4}: 'Ẵ'

LatinSmallLetterAWithBreveAndTilde

\u{1eb5}: 'ẵ'

LatinCapitalLetterAWithBreveAndDotBelow

\u{1eb6}: 'Ặ'

LatinSmallLetterAWithBreveAndDotBelow

\u{1eb7}: 'ặ'

LatinCapitalLetterEWithDotBelow

\u{1eb8}: 'Ẹ'

LatinSmallLetterEWithDotBelow

\u{1eb9}: 'ẹ'

LatinCapitalLetterEWithHookAbove

\u{1eba}: 'Ẻ'

LatinSmallLetterEWithHookAbove

\u{1ebb}: 'ẻ'

LatinCapitalLetterEWithTilde

\u{1ebc}: 'Ẽ'

LatinSmallLetterEWithTilde

\u{1ebd}: 'ẽ'

LatinCapitalLetterEWithCircumflexAndAcute

\u{1ebe}: 'Ế'

LatinSmallLetterEWithCircumflexAndAcute

\u{1ebf}: 'ế'

LatinCapitalLetterEWithCircumflexAndGrave

\u{1ec0}: 'Ề'

LatinSmallLetterEWithCircumflexAndGrave

\u{1ec1}: 'ề'

LatinCapitalLetterEWithCircumflexAndHookAbove

\u{1ec2}: 'Ể'

LatinSmallLetterEWithCircumflexAndHookAbove

\u{1ec3}: 'ể'

LatinCapitalLetterEWithCircumflexAndTilde

\u{1ec4}: 'Ễ'

LatinSmallLetterEWithCircumflexAndTilde

\u{1ec5}: 'ễ'

LatinCapitalLetterEWithCircumflexAndDotBelow

\u{1ec6}: 'Ệ'

LatinSmallLetterEWithCircumflexAndDotBelow

\u{1ec7}: 'ệ'

LatinCapitalLetterIWithHookAbove

\u{1ec8}: 'Ỉ'

LatinSmallLetterIWithHookAbove

\u{1ec9}: 'ỉ'

LatinCapitalLetterIWithDotBelow

\u{1eca}: 'Ị'

LatinSmallLetterIWithDotBelow

\u{1ecb}: 'ị'

LatinCapitalLetterOWithDotBelow

\u{1ecc}: 'Ọ'

LatinSmallLetterOWithDotBelow

\u{1ecd}: 'ọ'

LatinCapitalLetterOWithHookAbove

\u{1ece}: 'Ỏ'

LatinSmallLetterOWithHookAbove

\u{1ecf}: 'ỏ'

LatinCapitalLetterOWithCircumflexAndAcute

\u{1ed0}: 'Ố'

LatinSmallLetterOWithCircumflexAndAcute

\u{1ed1}: 'ố'

LatinCapitalLetterOWithCircumflexAndGrave

\u{1ed2}: 'Ồ'

LatinSmallLetterOWithCircumflexAndGrave

\u{1ed3}: 'ồ'

LatinCapitalLetterOWithCircumflexAndHookAbove

\u{1ed4}: 'Ổ'

LatinSmallLetterOWithCircumflexAndHookAbove

\u{1ed5}: 'ổ'

LatinCapitalLetterOWithCircumflexAndTilde

\u{1ed6}: 'Ỗ'

LatinSmallLetterOWithCircumflexAndTilde

\u{1ed7}: 'ỗ'

LatinCapitalLetterOWithCircumflexAndDotBelow

\u{1ed8}: 'Ộ'

LatinSmallLetterOWithCircumflexAndDotBelow

\u{1ed9}: 'ộ'

LatinCapitalLetterOWithHornAndAcute

\u{1eda}: 'Ớ'

LatinSmallLetterOWithHornAndAcute

\u{1edb}: 'ớ'

LatinCapitalLetterOWithHornAndGrave

\u{1edc}: 'Ờ'

LatinSmallLetterOWithHornAndGrave

\u{1edd}: 'ờ'

LatinCapitalLetterOWithHornAndHookAbove

\u{1ede}: 'Ở'

LatinSmallLetterOWithHornAndHookAbove

\u{1edf}: 'ở'

LatinCapitalLetterOWithHornAndTilde

\u{1ee0}: 'Ỡ'

LatinSmallLetterOWithHornAndTilde

\u{1ee1}: 'ỡ'

LatinCapitalLetterOWithHornAndDotBelow

\u{1ee2}: 'Ợ'

LatinSmallLetterOWithHornAndDotBelow

\u{1ee3}: 'ợ'

LatinCapitalLetterUWithDotBelow

\u{1ee4}: 'Ụ'

LatinSmallLetterUWithDotBelow

\u{1ee5}: 'ụ'

LatinCapitalLetterUWithHookAbove

\u{1ee6}: 'Ủ'

LatinSmallLetterUWithHookAbove

\u{1ee7}: 'ủ'

LatinCapitalLetterUWithHornAndAcute

\u{1ee8}: 'Ứ'

LatinSmallLetterUWithHornAndAcute

\u{1ee9}: 'ứ'

LatinCapitalLetterUWithHornAndGrave

\u{1eea}: 'Ừ'

LatinSmallLetterUWithHornAndGrave

\u{1eeb}: 'ừ'

LatinCapitalLetterUWithHornAndHookAbove

\u{1eec}: 'Ử'

LatinSmallLetterUWithHornAndHookAbove

\u{1eed}: 'ử'

LatinCapitalLetterUWithHornAndTilde

\u{1eee}: 'Ữ'

LatinSmallLetterUWithHornAndTilde

\u{1eef}: 'ữ'

LatinCapitalLetterUWithHornAndDotBelow

\u{1ef0}: 'Ự'

LatinSmallLetterUWithHornAndDotBelow

\u{1ef1}: 'ự'

LatinCapitalLetterYWithGrave

\u{1ef2}: 'Ỳ'

LatinSmallLetterYWithGrave

\u{1ef3}: 'ỳ'

LatinCapitalLetterYWithDotBelow

\u{1ef4}: 'Ỵ'

LatinSmallLetterYWithDotBelow

\u{1ef5}: 'ỵ'

LatinCapitalLetterYWithHookAbove

\u{1ef6}: 'Ỷ'

LatinSmallLetterYWithHookAbove

\u{1ef7}: 'ỷ'

LatinCapitalLetterYWithTilde

\u{1ef8}: 'Ỹ'

LatinSmallLetterYWithTilde

\u{1ef9}: 'ỹ'

LatinCapitalLetterMiddleDashWelshLl

\u{1efa}: 'Ỻ'

LatinSmallLetterMiddleDashWelshLl

\u{1efb}: 'ỻ'

LatinCapitalLetterMiddleDashWelshV

\u{1efc}: 'Ỽ'

LatinSmallLetterMiddleDashWelshV

\u{1efd}: 'ỽ'

LatinCapitalLetterYWithLoop

\u{1efe}: 'Ỿ'

Methods

impl LatinExtendedAdditional

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case