Enum unicode_types::latin_extended_b::LatinExtendedB

pub enum LatinExtendedB {
    LatinSmallLetterBWithStroke,
    LatinCapitalLetterBWithHook,
    LatinCapitalLetterBWithTopbar,
    LatinSmallLetterBWithTopbar,
    LatinCapitalLetterToneSix,
    LatinSmallLetterToneSix,
    LatinCapitalLetterOpenO,
    LatinCapitalLetterCWithHook,
    LatinSmallLetterCWithHook,
    LatinCapitalLetterAfricanD,
    LatinCapitalLetterDWithHook,
    LatinCapitalLetterDWithTopbar,
    LatinSmallLetterDWithTopbar,
    LatinSmallLetterTurnedDelta,
    LatinCapitalLetterReversedE,
    LatinCapitalLetterSchwa,
    LatinCapitalLetterOpenE,
    LatinCapitalLetterFWithHook,
    LatinSmallLetterFWithHook,
    LatinCapitalLetterGWithHook,
    LatinCapitalLetterGamma,
    LatinSmallLetterHv,
    LatinCapitalLetterIota,
    LatinCapitalLetterIWithStroke,
    LatinCapitalLetterKWithHook,
    LatinSmallLetterKWithHook,
    LatinSmallLetterLWithBar,
    LatinSmallLetterLambdaWithStroke,
    LatinCapitalLetterTurnedM,
    LatinCapitalLetterNWithLeftHook,
    LatinSmallLetterNWithLongRightLeg,
    LatinCapitalLetterOWithMiddleTilde,
    LatinCapitalLetterOWithHorn,
    LatinSmallLetterOWithHorn,
    LatinCapitalLetterOi,
    LatinSmallLetterOi,
    LatinCapitalLetterPWithHook,
    LatinSmallLetterPWithHook,
    LatinLetterYr,
    LatinCapitalLetterToneTwo,
    LatinSmallLetterToneTwo,
    LatinCapitalLetterEsh,
    LatinLetterReversedEshLoop,
    LatinSmallLetterTWithPalatalHook,
    LatinCapitalLetterTWithHook,
    LatinSmallLetterTWithHook,
    LatinCapitalLetterTWithRetroflexHook,
    LatinCapitalLetterUWithHorn,
    LatinSmallLetterUWithHorn,
    LatinCapitalLetterUpsilon,
    LatinCapitalLetterVWithHook,
    LatinCapitalLetterYWithHook,
    LatinSmallLetterYWithHook,
    LatinCapitalLetterZWithStroke,
    LatinSmallLetterZWithStroke,
    LatinCapitalLetterEzh,
    LatinCapitalLetterEzhReversed,
    LatinSmallLetterEzhReversed,
    LatinSmallLetterEzhWithTail,
    LatinLetterTwoWithStroke,
    LatinCapitalLetterToneFive,
    LatinSmallLetterToneFive,
    LatinLetterInvertedGlottalStopWithStroke,
    LatinLetterWynn,
    LatinLetterDentalClick,
    LatinLetterLateralClick,
    LatinLetterAlveolarClick,
    LatinLetterRetroflexClick,
    LatinCapitalLetterDzWithCaron,
    LatinCapitalLetterDWithSmallLetterZWithCaron,
    LatinSmallLetterDzWithCaron,
    LatinCapitalLetterLj,
    LatinCapitalLetterLWithSmallLetterJ,
    LatinSmallLetterLj,
    LatinCapitalLetterNj,
    LatinCapitalLetterNWithSmallLetterJ,
    LatinSmallLetterNj,
    LatinCapitalLetterAWithCaron,
    LatinSmallLetterAWithCaron,
    LatinCapitalLetterIWithCaron,
    LatinSmallLetterIWithCaron,
    LatinCapitalLetterOWithCaron,
    LatinSmallLetterOWithCaron,
    LatinCapitalLetterUWithCaron,
    LatinSmallLetterUWithCaron,
    LatinCapitalLetterUWithDiaeresisAndMacron,
    LatinSmallLetterUWithDiaeresisAndMacron,
    LatinCapitalLetterUWithDiaeresisAndAcute,
    LatinSmallLetterUWithDiaeresisAndAcute,
    LatinCapitalLetterUWithDiaeresisAndCaron,
    LatinSmallLetterUWithDiaeresisAndCaron,
    LatinCapitalLetterUWithDiaeresisAndGrave,
    LatinSmallLetterUWithDiaeresisAndGrave,
    LatinSmallLetterTurnedE,
    LatinCapitalLetterAWithDiaeresisAndMacron,
    LatinSmallLetterAWithDiaeresisAndMacron,
    LatinCapitalLetterAWithDotAboveAndMacron,
    LatinSmallLetterAWithDotAboveAndMacron,
    LatinCapitalLetterAeWithMacron,
    LatinSmallLetterAeWithMacron,
    LatinCapitalLetterGWithStroke,
    LatinSmallLetterGWithStroke,
    LatinCapitalLetterGWithCaron,
    LatinSmallLetterGWithCaron,
    LatinCapitalLetterKWithCaron,
    LatinSmallLetterKWithCaron,
    LatinCapitalLetterOWithOgonek,
    LatinSmallLetterOWithOgonek,
    LatinCapitalLetterOWithOgonekAndMacron,
    LatinSmallLetterOWithOgonekAndMacron,
    LatinCapitalLetterEzhWithCaron,
    LatinSmallLetterEzhWithCaron,
    LatinSmallLetterJWithCaron,
    LatinCapitalLetterDz,
    LatinCapitalLetterDWithSmallLetterZ,
    LatinSmallLetterDz,
    LatinCapitalLetterGWithAcute,
    LatinSmallLetterGWithAcute,
    LatinCapitalLetterHwair,
    LatinCapitalLetterWynn,
    LatinCapitalLetterNWithGrave,
    LatinSmallLetterNWithGrave,
    LatinCapitalLetterAWithRingAboveAndAcute,
    LatinSmallLetterAWithRingAboveAndAcute,
    LatinCapitalLetterAeWithAcute,
    LatinSmallLetterAeWithAcute,
    LatinCapitalLetterOWithStrokeAndAcute,
    LatinSmallLetterOWithStrokeAndAcute,
    LatinCapitalLetterAWithDoubleGrave,
    LatinSmallLetterAWithDoubleGrave,
    LatinCapitalLetterAWithInvertedBreve,
    LatinSmallLetterAWithInvertedBreve,
    LatinCapitalLetterEWithDoubleGrave,
    LatinSmallLetterEWithDoubleGrave,
    LatinCapitalLetterEWithInvertedBreve,
    LatinSmallLetterEWithInvertedBreve,
    LatinCapitalLetterIWithDoubleGrave,
    LatinSmallLetterIWithDoubleGrave,
    LatinCapitalLetterIWithInvertedBreve,
    LatinSmallLetterIWithInvertedBreve,
    LatinCapitalLetterOWithDoubleGrave,
    LatinSmallLetterOWithDoubleGrave,
    LatinCapitalLetterOWithInvertedBreve,
    LatinSmallLetterOWithInvertedBreve,
    LatinCapitalLetterRWithDoubleGrave,
    LatinSmallLetterRWithDoubleGrave,
    LatinCapitalLetterRWithInvertedBreve,
    LatinSmallLetterRWithInvertedBreve,
    LatinCapitalLetterUWithDoubleGrave,
    LatinSmallLetterUWithDoubleGrave,
    LatinCapitalLetterUWithInvertedBreve,
    LatinSmallLetterUWithInvertedBreve,
    LatinCapitalLetterSWithCommaBelow,
    LatinSmallLetterSWithCommaBelow,
    LatinCapitalLetterTWithCommaBelow,
    LatinSmallLetterTWithCommaBelow,
    LatinCapitalLetterYogh,
    LatinSmallLetterYogh,
    LatinCapitalLetterHWithCaron,
    LatinSmallLetterHWithCaron,
    LatinCapitalLetterNWithLongRightLeg,
    LatinSmallLetterDWithCurl,
    LatinCapitalLetterOu,
    LatinSmallLetterOu,
    LatinCapitalLetterZWithHook,
    LatinSmallLetterZWithHook,
    LatinCapitalLetterAWithDotAbove,
    LatinSmallLetterAWithDotAbove,
    LatinCapitalLetterEWithCedilla,
    LatinSmallLetterEWithCedilla,
    LatinCapitalLetterOWithDiaeresisAndMacron,
    LatinSmallLetterOWithDiaeresisAndMacron,
    LatinCapitalLetterOWithTildeAndMacron,
    LatinSmallLetterOWithTildeAndMacron,
    LatinCapitalLetterOWithDotAbove,
    LatinSmallLetterOWithDotAbove,
    LatinCapitalLetterOWithDotAboveAndMacron,
    LatinSmallLetterOWithDotAboveAndMacron,
    LatinCapitalLetterYWithMacron,
    LatinSmallLetterYWithMacron,
    LatinSmallLetterLWithCurl,
    LatinSmallLetterNWithCurl,
    LatinSmallLetterTWithCurl,
    LatinSmallLetterDotlessJ,
    LatinSmallLetterDbDigraph,
    LatinSmallLetterQpDigraph,
    LatinCapitalLetterAWithStroke,
    LatinCapitalLetterCWithStroke,
    LatinSmallLetterCWithStroke,
    LatinCapitalLetterLWithBar,
    LatinCapitalLetterTWithDiagonalStroke,
    LatinSmallLetterSWithSwashTail,
    LatinSmallLetterZWithSwashTail,
    LatinCapitalLetterGlottalStop,
    LatinSmallLetterGlottalStop,
    LatinCapitalLetterBWithStroke,
    LatinCapitalLetterUBar,
    LatinCapitalLetterTurnedV,
    LatinCapitalLetterEWithStroke,
    LatinSmallLetterEWithStroke,
    LatinCapitalLetterJWithStroke,
    LatinSmallLetterJWithStroke,
    LatinCapitalLetterSmallQWithHookTail,
    LatinSmallLetterQWithHookTail,
    LatinCapitalLetterRWithStroke,
    LatinSmallLetterRWithStroke,
    LatinCapitalLetterYWithStroke,
}

An enum to represent all characters in the LatinExtendedB block.

Variants
LatinSmallLetterBWithStroke

\u{180}: 'ƀ'

LatinCapitalLetterBWithHook

\u{181}: 'Ɓ'

LatinCapitalLetterBWithTopbar

\u{182}: 'Ƃ'

LatinSmallLetterBWithTopbar

\u{183}: 'ƃ'

LatinCapitalLetterToneSix

\u{184}: 'Ƅ'

LatinSmallLetterToneSix

\u{185}: 'ƅ'

LatinCapitalLetterOpenO

\u{186}: 'Ɔ'

LatinCapitalLetterCWithHook

\u{187}: 'Ƈ'

LatinSmallLetterCWithHook

\u{188}: 'ƈ'

LatinCapitalLetterAfricanD

\u{189}: 'Ɖ'

LatinCapitalLetterDWithHook

\u{18a}: 'Ɗ'

LatinCapitalLetterDWithTopbar

\u{18b}: 'Ƌ'

LatinSmallLetterDWithTopbar

\u{18c}: 'ƌ'

LatinSmallLetterTurnedDelta

\u{18d}: 'ƍ'

LatinCapitalLetterReversedE

\u{18e}: 'Ǝ'

LatinCapitalLetterSchwa

\u{18f}: 'Ə'

LatinCapitalLetterOpenE

\u{190}: 'Ɛ'

LatinCapitalLetterFWithHook

\u{191}: 'Ƒ'

LatinSmallLetterFWithHook

\u{192}: 'ƒ'

LatinCapitalLetterGWithHook

\u{193}: 'Ɠ'

LatinCapitalLetterGamma

\u{194}: 'Ɣ'

LatinSmallLetterHv

\u{195}: 'ƕ'

LatinCapitalLetterIota

\u{196}: 'Ɩ'

LatinCapitalLetterIWithStroke

\u{197}: 'Ɨ'

LatinCapitalLetterKWithHook

\u{198}: 'Ƙ'

LatinSmallLetterKWithHook

\u{199}: 'ƙ'

LatinSmallLetterLWithBar

\u{19a}: 'ƚ'

LatinSmallLetterLambdaWithStroke

\u{19b}: 'ƛ'

LatinCapitalLetterTurnedM

\u{19c}: 'Ɯ'

LatinCapitalLetterNWithLeftHook

\u{19d}: 'Ɲ'

LatinSmallLetterNWithLongRightLeg

\u{19e}: 'ƞ'

LatinCapitalLetterOWithMiddleTilde

\u{19f}: 'Ɵ'

LatinCapitalLetterOWithHorn

\u{1a0}: 'Ơ'

LatinSmallLetterOWithHorn

\u{1a1}: 'ơ'

LatinCapitalLetterOi

\u{1a2}: 'Ƣ'

LatinSmallLetterOi

\u{1a3}: 'ƣ'

LatinCapitalLetterPWithHook

\u{1a4}: 'Ƥ'

LatinSmallLetterPWithHook

\u{1a5}: 'ƥ'

LatinLetterYr

\u{1a6}: 'Ʀ'

LatinCapitalLetterToneTwo

\u{1a7}: 'Ƨ'

LatinSmallLetterToneTwo

\u{1a8}: 'ƨ'

LatinCapitalLetterEsh

\u{1a9}: 'Ʃ'

LatinLetterReversedEshLoop

\u{1aa}: 'ƪ'

LatinSmallLetterTWithPalatalHook

\u{1ab}: 'ƫ'

LatinCapitalLetterTWithHook

\u{1ac}: 'Ƭ'

LatinSmallLetterTWithHook

\u{1ad}: 'ƭ'

LatinCapitalLetterTWithRetroflexHook

\u{1ae}: 'Ʈ'

LatinCapitalLetterUWithHorn

\u{1af}: 'Ư'

LatinSmallLetterUWithHorn

\u{1b0}: 'ư'

LatinCapitalLetterUpsilon

\u{1b1}: 'Ʊ'

LatinCapitalLetterVWithHook

\u{1b2}: 'Ʋ'

LatinCapitalLetterYWithHook

\u{1b3}: 'Ƴ'

LatinSmallLetterYWithHook

\u{1b4}: 'ƴ'

LatinCapitalLetterZWithStroke

\u{1b5}: 'Ƶ'

LatinSmallLetterZWithStroke

\u{1b6}: 'ƶ'

LatinCapitalLetterEzh

\u{1b7}: 'Ʒ'

LatinCapitalLetterEzhReversed

\u{1b8}: 'Ƹ'

LatinSmallLetterEzhReversed

\u{1b9}: 'ƹ'

LatinSmallLetterEzhWithTail

\u{1ba}: 'ƺ'

LatinLetterTwoWithStroke

\u{1bb}: 'ƻ'

LatinCapitalLetterToneFive

\u{1bc}: 'Ƽ'

LatinSmallLetterToneFive

\u{1bd}: 'ƽ'

LatinLetterInvertedGlottalStopWithStroke

\u{1be}: 'ƾ'

LatinLetterWynn

\u{1bf}: 'ƿ'

LatinLetterDentalClick

\u{1c0}: 'ǀ'

LatinLetterLateralClick

\u{1c1}: 'ǁ'

LatinLetterAlveolarClick

\u{1c2}: 'ǂ'

LatinLetterRetroflexClick

\u{1c3}: 'ǃ'

LatinCapitalLetterDzWithCaron

\u{1c4}: 'Ǆ'

LatinCapitalLetterDWithSmallLetterZWithCaron

\u{1c5}: 'ǅ'

LatinSmallLetterDzWithCaron

\u{1c6}: 'ǆ'

LatinCapitalLetterLj

\u{1c7}: 'Ǉ'

LatinCapitalLetterLWithSmallLetterJ

\u{1c8}: 'ǈ'

LatinSmallLetterLj

\u{1c9}: 'ǉ'

LatinCapitalLetterNj

\u{1ca}: 'Ǌ'

LatinCapitalLetterNWithSmallLetterJ

\u{1cb}: 'ǋ'

LatinSmallLetterNj

\u{1cc}: 'ǌ'

LatinCapitalLetterAWithCaron

\u{1cd}: 'Ǎ'

LatinSmallLetterAWithCaron

\u{1ce}: 'ǎ'

LatinCapitalLetterIWithCaron

\u{1cf}: 'Ǐ'

LatinSmallLetterIWithCaron

\u{1d0}: 'ǐ'

LatinCapitalLetterOWithCaron

\u{1d1}: 'Ǒ'

LatinSmallLetterOWithCaron

\u{1d2}: 'ǒ'

LatinCapitalLetterUWithCaron

\u{1d3}: 'Ǔ'

LatinSmallLetterUWithCaron

\u{1d4}: 'ǔ'

LatinCapitalLetterUWithDiaeresisAndMacron

\u{1d5}: 'Ǖ'

LatinSmallLetterUWithDiaeresisAndMacron

\u{1d6}: 'ǖ'

LatinCapitalLetterUWithDiaeresisAndAcute

\u{1d7}: 'Ǘ'

LatinSmallLetterUWithDiaeresisAndAcute

\u{1d8}: 'ǘ'

LatinCapitalLetterUWithDiaeresisAndCaron

\u{1d9}: 'Ǚ'

LatinSmallLetterUWithDiaeresisAndCaron

\u{1da}: 'ǚ'

LatinCapitalLetterUWithDiaeresisAndGrave

\u{1db}: 'Ǜ'

LatinSmallLetterUWithDiaeresisAndGrave

\u{1dc}: 'ǜ'

LatinSmallLetterTurnedE

\u{1dd}: 'ǝ'

LatinCapitalLetterAWithDiaeresisAndMacron

\u{1de}: 'Ǟ'

LatinSmallLetterAWithDiaeresisAndMacron

\u{1df}: 'ǟ'

LatinCapitalLetterAWithDotAboveAndMacron

\u{1e0}: 'Ǡ'

LatinSmallLetterAWithDotAboveAndMacron

\u{1e1}: 'ǡ'

LatinCapitalLetterAeWithMacron

\u{1e2}: 'Ǣ'

LatinSmallLetterAeWithMacron

\u{1e3}: 'ǣ'

LatinCapitalLetterGWithStroke

\u{1e4}: 'Ǥ'

LatinSmallLetterGWithStroke

\u{1e5}: 'ǥ'

LatinCapitalLetterGWithCaron

\u{1e6}: 'Ǧ'

LatinSmallLetterGWithCaron

\u{1e7}: 'ǧ'

LatinCapitalLetterKWithCaron

\u{1e8}: 'Ǩ'

LatinSmallLetterKWithCaron

\u{1e9}: 'ǩ'

LatinCapitalLetterOWithOgonek

\u{1ea}: 'Ǫ'

LatinSmallLetterOWithOgonek

\u{1eb}: 'ǫ'

LatinCapitalLetterOWithOgonekAndMacron

\u{1ec}: 'Ǭ'

LatinSmallLetterOWithOgonekAndMacron

\u{1ed}: 'ǭ'

LatinCapitalLetterEzhWithCaron

\u{1ee}: 'Ǯ'

LatinSmallLetterEzhWithCaron

\u{1ef}: 'ǯ'

LatinSmallLetterJWithCaron

\u{1f0}: 'ǰ'

LatinCapitalLetterDz

\u{1f1}: 'Ǳ'

LatinCapitalLetterDWithSmallLetterZ

\u{1f2}: 'ǲ'

LatinSmallLetterDz

\u{1f3}: 'ǳ'

LatinCapitalLetterGWithAcute

\u{1f4}: 'Ǵ'

LatinSmallLetterGWithAcute

\u{1f5}: 'ǵ'

LatinCapitalLetterHwair

\u{1f6}: 'Ƕ'

LatinCapitalLetterWynn

\u{1f7}: 'Ƿ'

LatinCapitalLetterNWithGrave

\u{1f8}: 'Ǹ'

LatinSmallLetterNWithGrave

\u{1f9}: 'ǹ'

LatinCapitalLetterAWithRingAboveAndAcute

\u{1fa}: 'Ǻ'

LatinSmallLetterAWithRingAboveAndAcute

\u{1fb}: 'ǻ'

LatinCapitalLetterAeWithAcute

\u{1fc}: 'Ǽ'

LatinSmallLetterAeWithAcute

\u{1fd}: 'ǽ'

LatinCapitalLetterOWithStrokeAndAcute

\u{1fe}: 'Ǿ'

LatinSmallLetterOWithStrokeAndAcute

\u{1ff}: 'ǿ'

LatinCapitalLetterAWithDoubleGrave

\u{200}: 'Ȁ'

LatinSmallLetterAWithDoubleGrave

\u{201}: 'ȁ'

LatinCapitalLetterAWithInvertedBreve

\u{202}: 'Ȃ'

LatinSmallLetterAWithInvertedBreve

\u{203}: 'ȃ'

LatinCapitalLetterEWithDoubleGrave

\u{204}: 'Ȅ'

LatinSmallLetterEWithDoubleGrave

\u{205}: 'ȅ'

LatinCapitalLetterEWithInvertedBreve

\u{206}: 'Ȇ'

LatinSmallLetterEWithInvertedBreve

\u{207}: 'ȇ'

LatinCapitalLetterIWithDoubleGrave

\u{208}: 'Ȉ'

LatinSmallLetterIWithDoubleGrave

\u{209}: 'ȉ'

LatinCapitalLetterIWithInvertedBreve

\u{20a}: 'Ȋ'

LatinSmallLetterIWithInvertedBreve

\u{20b}: 'ȋ'

LatinCapitalLetterOWithDoubleGrave

\u{20c}: 'Ȍ'

LatinSmallLetterOWithDoubleGrave

\u{20d}: 'ȍ'

LatinCapitalLetterOWithInvertedBreve

\u{20e}: 'Ȏ'

LatinSmallLetterOWithInvertedBreve

\u{20f}: 'ȏ'

LatinCapitalLetterRWithDoubleGrave

\u{210}: 'Ȑ'

LatinSmallLetterRWithDoubleGrave

\u{211}: 'ȑ'

LatinCapitalLetterRWithInvertedBreve

\u{212}: 'Ȓ'

LatinSmallLetterRWithInvertedBreve

\u{213}: 'ȓ'

LatinCapitalLetterUWithDoubleGrave

\u{214}: 'Ȕ'

LatinSmallLetterUWithDoubleGrave

\u{215}: 'ȕ'

LatinCapitalLetterUWithInvertedBreve

\u{216}: 'Ȗ'

LatinSmallLetterUWithInvertedBreve

\u{217}: 'ȗ'

LatinCapitalLetterSWithCommaBelow

\u{218}: 'Ș'

LatinSmallLetterSWithCommaBelow

\u{219}: 'ș'

LatinCapitalLetterTWithCommaBelow

\u{21a}: 'Ț'

LatinSmallLetterTWithCommaBelow

\u{21b}: 'ț'

LatinCapitalLetterYogh

\u{21c}: 'Ȝ'

LatinSmallLetterYogh

\u{21d}: 'ȝ'

LatinCapitalLetterHWithCaron

\u{21e}: 'Ȟ'

LatinSmallLetterHWithCaron

\u{21f}: 'ȟ'

LatinCapitalLetterNWithLongRightLeg

\u{220}: 'Ƞ'

LatinSmallLetterDWithCurl

\u{221}: 'ȡ'

LatinCapitalLetterOu

\u{222}: 'Ȣ'

LatinSmallLetterOu

\u{223}: 'ȣ'

LatinCapitalLetterZWithHook

\u{224}: 'Ȥ'

LatinSmallLetterZWithHook

\u{225}: 'ȥ'

LatinCapitalLetterAWithDotAbove

\u{226}: 'Ȧ'

LatinSmallLetterAWithDotAbove

\u{227}: 'ȧ'

LatinCapitalLetterEWithCedilla

\u{228}: 'Ȩ'

LatinSmallLetterEWithCedilla

\u{229}: 'ȩ'

LatinCapitalLetterOWithDiaeresisAndMacron

\u{22a}: 'Ȫ'

LatinSmallLetterOWithDiaeresisAndMacron

\u{22b}: 'ȫ'

LatinCapitalLetterOWithTildeAndMacron

\u{22c}: 'Ȭ'

LatinSmallLetterOWithTildeAndMacron

\u{22d}: 'ȭ'

LatinCapitalLetterOWithDotAbove

\u{22e}: 'Ȯ'

LatinSmallLetterOWithDotAbove

\u{22f}: 'ȯ'

LatinCapitalLetterOWithDotAboveAndMacron

\u{230}: 'Ȱ'

LatinSmallLetterOWithDotAboveAndMacron

\u{231}: 'ȱ'

LatinCapitalLetterYWithMacron

\u{232}: 'Ȳ'

LatinSmallLetterYWithMacron

\u{233}: 'ȳ'

LatinSmallLetterLWithCurl

\u{234}: 'ȴ'

LatinSmallLetterNWithCurl

\u{235}: 'ȵ'

LatinSmallLetterTWithCurl

\u{236}: 'ȶ'

LatinSmallLetterDotlessJ

\u{237}: 'ȷ'

LatinSmallLetterDbDigraph

\u{238}: 'ȸ'

LatinSmallLetterQpDigraph

\u{239}: 'ȹ'

LatinCapitalLetterAWithStroke

\u{23a}: 'Ⱥ'

LatinCapitalLetterCWithStroke

\u{23b}: 'Ȼ'

LatinSmallLetterCWithStroke

\u{23c}: 'ȼ'

LatinCapitalLetterLWithBar

\u{23d}: 'Ƚ'

LatinCapitalLetterTWithDiagonalStroke

\u{23e}: 'Ⱦ'

LatinSmallLetterSWithSwashTail

\u{23f}: 'ȿ'

LatinSmallLetterZWithSwashTail

\u{240}: 'ɀ'

LatinCapitalLetterGlottalStop

\u{241}: 'Ɂ'

LatinSmallLetterGlottalStop

\u{242}: 'ɂ'

LatinCapitalLetterBWithStroke

\u{243}: 'Ƀ'

LatinCapitalLetterUBar

\u{244}: 'Ʉ'

LatinCapitalLetterTurnedV

\u{245}: 'Ʌ'

LatinCapitalLetterEWithStroke

\u{246}: 'Ɇ'

LatinSmallLetterEWithStroke

\u{247}: 'ɇ'

LatinCapitalLetterJWithStroke

\u{248}: 'Ɉ'

LatinSmallLetterJWithStroke

\u{249}: 'ɉ'

LatinCapitalLetterSmallQWithHookTail

\u{24a}: 'Ɋ'

LatinSmallLetterQWithHookTail

\u{24b}: 'ɋ'

LatinCapitalLetterRWithStroke

\u{24c}: 'Ɍ'

LatinSmallLetterRWithStroke

\u{24d}: 'ɍ'

LatinCapitalLetterYWithStroke

\u{24e}: 'Ɏ'

Methods

impl LatinExtendedB

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case