Enum unicode_types::latin_extended_d::LatinExtendedD

pub enum LatinExtendedD {
    ModifierLetterStressAndHighTone,
    ModifierLetterStressAndLowTone,
    LatinCapitalLetterEgyptologicalAlef,
    LatinSmallLetterEgyptologicalAlef,
    LatinCapitalLetterEgyptologicalAin,
    LatinSmallLetterEgyptologicalAin,
    LatinCapitalLetterHeng,
    LatinSmallLetterHeng,
    LatinCapitalLetterTz,
    LatinSmallLetterTz,
    LatinCapitalLetterTresillo,
    LatinSmallLetterTresillo,
    LatinCapitalLetterCuatrillo,
    LatinSmallLetterCuatrillo,
    LatinCapitalLetterCuatrilloWithComma,
    LatinSmallLetterCuatrilloWithComma,
    LatinLetterSmallCapitalF,
    LatinLetterSmallCapitalS,
    LatinCapitalLetterAa,
    LatinSmallLetterAa,
    LatinCapitalLetterAo,
    LatinSmallLetterAo,
    LatinCapitalLetterAu,
    LatinSmallLetterAu,
    LatinCapitalLetterAv,
    LatinSmallLetterAv,
    LatinCapitalLetterAvWithHorizontalBar,
    LatinSmallLetterAvWithHorizontalBar,
    LatinCapitalLetterAy,
    LatinSmallLetterAy,
    LatinCapitalLetterReversedCWithDot,
    LatinSmallLetterReversedCWithDot,
    LatinCapitalLetterKWithStroke,
    LatinSmallLetterKWithStroke,
    LatinCapitalLetterKWithDiagonalStroke,
    LatinSmallLetterKWithDiagonalStroke,
    LatinCapitalLetterKWithStrokeAndDiagonalStroke,
    LatinSmallLetterKWithStrokeAndDiagonalStroke,
    LatinCapitalLetterBrokenL,
    LatinSmallLetterBrokenL,
    LatinCapitalLetterLWithHighStroke,
    LatinSmallLetterLWithHighStroke,
    LatinCapitalLetterOWithLongStrokeOverlay,
    LatinSmallLetterOWithLongStrokeOverlay,
    LatinCapitalLetterOWithLoop,
    LatinSmallLetterOWithLoop,
    LatinCapitalLetterOo,
    LatinSmallLetterOo,
    LatinCapitalLetterPWithStrokeThroughDescender,
    LatinSmallLetterPWithStrokeThroughDescender,
    LatinCapitalLetterPWithFlourish,
    LatinSmallLetterPWithFlourish,
    LatinCapitalLetterPWithSquirrelTail,
    LatinSmallLetterPWithSquirrelTail,
    LatinCapitalLetterQWithStrokeThroughDescender,
    LatinSmallLetterQWithStrokeThroughDescender,
    LatinCapitalLetterQWithDiagonalStroke,
    LatinSmallLetterQWithDiagonalStroke,
    LatinCapitalLetterRRotunda,
    LatinSmallLetterRRotunda,
    LatinCapitalLetterRumRotunda,
    LatinSmallLetterRumRotunda,
    LatinCapitalLetterVWithDiagonalStroke,
    LatinSmallLetterVWithDiagonalStroke,
    LatinCapitalLetterVy,
    LatinSmallLetterVy,
    LatinCapitalLetterVisigothicZ,
    LatinSmallLetterVisigothicZ,
    LatinCapitalLetterThornWithStroke,
    LatinSmallLetterThornWithStroke,
    LatinCapitalLetterThornWithStrokeThroughDescender,
    LatinSmallLetterThornWithStrokeThroughDescender,
    LatinCapitalLetterVend,
    LatinSmallLetterVend,
    LatinCapitalLetterEt,
    LatinSmallLetterEt,
    LatinCapitalLetterIs,
    LatinSmallLetterIs,
    LatinCapitalLetterCon,
    LatinSmallLetterCon,
    ModifierLetterUs,
    LatinSmallLetterDum,
    LatinSmallLetterLum,
    LatinSmallLetterMum,
    LatinSmallLetterNum,
    LatinSmallLetterRum,
    LatinLetterSmallCapitalRum,
    LatinSmallLetterTum,
    LatinSmallLetterUm,
    LatinCapitalLetterInsularD,
    LatinSmallLetterInsularD,
    LatinCapitalLetterInsularF,
    LatinSmallLetterInsularF,
    LatinCapitalLetterInsularG,
    LatinCapitalLetterTurnedInsularG,
    LatinSmallLetterTurnedInsularG,
    LatinCapitalLetterTurnedL,
    LatinSmallLetterTurnedL,
    LatinCapitalLetterInsularR,
    LatinSmallLetterInsularR,
    LatinCapitalLetterInsularS,
    LatinSmallLetterInsularS,
    LatinCapitalLetterInsularT,
    LatinSmallLetterInsularT,
    ModifierLetterLowCircumflexAccent,
    ModifierLetterColon,
    ModifierLetterShortEqualsSign,
    LatinCapitalLetterSaltillo,
    LatinSmallLetterSaltillo,
    LatinCapitalLetterTurnedH,
    LatinSmallLetterLWithRetroflexHookAndBelt,
    LatinLetterSinologicalDot,
    LatinCapitalLetterNWithDescender,
    LatinSmallLetterNWithDescender,
    LatinCapitalLetterCWithBar,
    LatinSmallLetterCWithBar,
    LatinSmallLetterCWithPalatalHook,
    LatinSmallLetterHWithPalatalHook,
    LatinCapitalLetterBWithFlourish,
    LatinSmallLetterBWithFlourish,
    LatinCapitalLetterFWithStroke,
    LatinSmallLetterFWithStroke,
    LatinCapitalLetterVolapukAe,
    LatinSmallLetterVolapukAe,
    LatinCapitalLetterVolapukOe,
    LatinSmallLetterVolapukOe,
    LatinCapitalLetterVolapukUe,
    LatinSmallLetterVolapukUe,
    LatinCapitalLetterGWithObliqueStroke,
    LatinSmallLetterGWithObliqueStroke,
    LatinCapitalLetterKWithObliqueStroke,
    LatinSmallLetterKWithObliqueStroke,
    LatinCapitalLetterNWithObliqueStroke,
    LatinSmallLetterNWithObliqueStroke,
    LatinCapitalLetterRWithObliqueStroke,
    LatinSmallLetterRWithObliqueStroke,
    LatinCapitalLetterSWithObliqueStroke,
    LatinSmallLetterSWithObliqueStroke,
    LatinCapitalLetterHWithHook,
    LatinCapitalLetterReversedOpenE,
    LatinCapitalLetterScriptG,
    LatinCapitalLetterLWithBelt,
    LatinCapitalLetterSmallCapitalI,
    LatinLetterSmallCapitalQ,
    LatinCapitalLetterTurnedK,
    LatinCapitalLetterTurnedT,
    LatinCapitalLetterJWithCrossedDashTail,
    LatinCapitalLetterChi,
    LatinCapitalLetterBeta,
    LatinSmallLetterBeta,
    LatinCapitalLetterOmega,
    LatinSmallLetterOmega,
    LatinCapitalLetterUWithStroke,
    LatinSmallLetterUWithStroke,
    LatinCapitalLetterGlottalA,
    LatinSmallLetterGlottalA,
    LatinCapitalLetterGlottalI,
    LatinSmallLetterGlottalI,
    LatinCapitalLetterGlottalU,
    LatinSmallLetterGlottalU,
    LatinCapitalLetterAnglicanaW,
    LatinSmallLetterAnglicanaW,
    LatinCapitalLetterCWithPalatalHook,
    LatinCapitalLetterSWithHook,
    LatinCapitalLetterZWithPalatalHook,
    LatinEpigraphicLetterSidewaysI,
    ModifierLetterCapitalHWithStroke,
    ModifierLetterSmallLigatureOe,
    LatinLetterSmallCapitalTurnedM,
    LatinEpigraphicLetterReversedF,
    LatinEpigraphicLetterReversedP,
    LatinEpigraphicLetterInvertedM,
    LatinEpigraphicLetterILonga,
}

An enum to represent all characters in the LatinExtendedD block.

Variants
ModifierLetterStressAndHighTone

\u{a720}: '꜠'

ModifierLetterStressAndLowTone

\u{a721}: '꜡'

LatinCapitalLetterEgyptologicalAlef

\u{a722}: 'Ꜣ'

LatinSmallLetterEgyptologicalAlef

\u{a723}: 'ꜣ'

LatinCapitalLetterEgyptologicalAin

\u{a724}: 'Ꜥ'

LatinSmallLetterEgyptologicalAin

\u{a725}: 'ꜥ'

LatinCapitalLetterHeng

\u{a726}: 'Ꜧ'

LatinSmallLetterHeng

\u{a727}: 'ꜧ'

LatinCapitalLetterTz

\u{a728}: 'Ꜩ'

LatinSmallLetterTz

\u{a729}: 'ꜩ'

LatinCapitalLetterTresillo

\u{a72a}: 'Ꜫ'

LatinSmallLetterTresillo

\u{a72b}: 'ꜫ'

LatinCapitalLetterCuatrillo

\u{a72c}: 'Ꜭ'

LatinSmallLetterCuatrillo

\u{a72d}: 'ꜭ'

LatinCapitalLetterCuatrilloWithComma

\u{a72e}: 'Ꜯ'

LatinSmallLetterCuatrilloWithComma

\u{a72f}: 'ꜯ'

LatinLetterSmallCapitalF

\u{a730}: 'ꜰ'

LatinLetterSmallCapitalS

\u{a731}: 'ꜱ'

LatinCapitalLetterAa

\u{a732}: 'Ꜳ'

LatinSmallLetterAa

\u{a733}: 'ꜳ'

LatinCapitalLetterAo

\u{a734}: 'Ꜵ'

LatinSmallLetterAo

\u{a735}: 'ꜵ'

LatinCapitalLetterAu

\u{a736}: 'Ꜷ'

LatinSmallLetterAu

\u{a737}: 'ꜷ'

LatinCapitalLetterAv

\u{a738}: 'Ꜹ'

LatinSmallLetterAv

\u{a739}: 'ꜹ'

LatinCapitalLetterAvWithHorizontalBar

\u{a73a}: 'Ꜻ'

LatinSmallLetterAvWithHorizontalBar

\u{a73b}: 'ꜻ'

LatinCapitalLetterAy

\u{a73c}: 'Ꜽ'

LatinSmallLetterAy

\u{a73d}: 'ꜽ'

LatinCapitalLetterReversedCWithDot

\u{a73e}: 'Ꜿ'

LatinSmallLetterReversedCWithDot

\u{a73f}: 'ꜿ'

LatinCapitalLetterKWithStroke

\u{a740}: 'Ꝁ'

LatinSmallLetterKWithStroke

\u{a741}: 'ꝁ'

LatinCapitalLetterKWithDiagonalStroke

\u{a742}: 'Ꝃ'

LatinSmallLetterKWithDiagonalStroke

\u{a743}: 'ꝃ'

LatinCapitalLetterKWithStrokeAndDiagonalStroke

\u{a744}: 'Ꝅ'

LatinSmallLetterKWithStrokeAndDiagonalStroke

\u{a745}: 'ꝅ'

LatinCapitalLetterBrokenL

\u{a746}: 'Ꝇ'

LatinSmallLetterBrokenL

\u{a747}: 'ꝇ'

LatinCapitalLetterLWithHighStroke

\u{a748}: 'Ꝉ'

LatinSmallLetterLWithHighStroke

\u{a749}: 'ꝉ'

LatinCapitalLetterOWithLongStrokeOverlay

\u{a74a}: 'Ꝋ'

LatinSmallLetterOWithLongStrokeOverlay

\u{a74b}: 'ꝋ'

LatinCapitalLetterOWithLoop

\u{a74c}: 'Ꝍ'

LatinSmallLetterOWithLoop

\u{a74d}: 'ꝍ'

LatinCapitalLetterOo

\u{a74e}: 'Ꝏ'

LatinSmallLetterOo

\u{a74f}: 'ꝏ'

LatinCapitalLetterPWithStrokeThroughDescender

\u{a750}: 'Ꝑ'

LatinSmallLetterPWithStrokeThroughDescender

\u{a751}: 'ꝑ'

LatinCapitalLetterPWithFlourish

\u{a752}: 'Ꝓ'

LatinSmallLetterPWithFlourish

\u{a753}: 'ꝓ'

LatinCapitalLetterPWithSquirrelTail

\u{a754}: 'Ꝕ'

LatinSmallLetterPWithSquirrelTail

\u{a755}: 'ꝕ'

LatinCapitalLetterQWithStrokeThroughDescender

\u{a756}: 'Ꝗ'

LatinSmallLetterQWithStrokeThroughDescender

\u{a757}: 'ꝗ'

LatinCapitalLetterQWithDiagonalStroke

\u{a758}: 'Ꝙ'

LatinSmallLetterQWithDiagonalStroke

\u{a759}: 'ꝙ'

LatinCapitalLetterRRotunda

\u{a75a}: 'Ꝛ'

LatinSmallLetterRRotunda

\u{a75b}: 'ꝛ'

LatinCapitalLetterRumRotunda

\u{a75c}: 'Ꝝ'

LatinSmallLetterRumRotunda

\u{a75d}: 'ꝝ'

LatinCapitalLetterVWithDiagonalStroke

\u{a75e}: 'Ꝟ'

LatinSmallLetterVWithDiagonalStroke

\u{a75f}: 'ꝟ'

LatinCapitalLetterVy

\u{a760}: 'Ꝡ'

LatinSmallLetterVy

\u{a761}: 'ꝡ'

LatinCapitalLetterVisigothicZ

\u{a762}: 'Ꝣ'

LatinSmallLetterVisigothicZ

\u{a763}: 'ꝣ'

LatinCapitalLetterThornWithStroke

\u{a764}: 'Ꝥ'

LatinSmallLetterThornWithStroke

\u{a765}: 'ꝥ'

LatinCapitalLetterThornWithStrokeThroughDescender

\u{a766}: 'Ꝧ'

LatinSmallLetterThornWithStrokeThroughDescender

\u{a767}: 'ꝧ'

LatinCapitalLetterVend

\u{a768}: 'Ꝩ'

LatinSmallLetterVend

\u{a769}: 'ꝩ'

LatinCapitalLetterEt

\u{a76a}: 'Ꝫ'

LatinSmallLetterEt

\u{a76b}: 'ꝫ'

LatinCapitalLetterIs

\u{a76c}: 'Ꝭ'

LatinSmallLetterIs

\u{a76d}: 'ꝭ'

LatinCapitalLetterCon

\u{a76e}: 'Ꝯ'

LatinSmallLetterCon

\u{a76f}: 'ꝯ'

ModifierLetterUs

\u{a770}: 'ꝰ'

LatinSmallLetterDum

\u{a771}: 'ꝱ'

LatinSmallLetterLum

\u{a772}: 'ꝲ'

LatinSmallLetterMum

\u{a773}: 'ꝳ'

LatinSmallLetterNum

\u{a774}: 'ꝴ'

LatinSmallLetterRum

\u{a775}: 'ꝵ'

LatinLetterSmallCapitalRum

\u{a776}: 'ꝶ'

LatinSmallLetterTum

\u{a777}: 'ꝷ'

LatinSmallLetterUm

\u{a778}: 'ꝸ'

LatinCapitalLetterInsularD

\u{a779}: 'Ꝺ'

LatinSmallLetterInsularD

\u{a77a}: 'ꝺ'

LatinCapitalLetterInsularF

\u{a77b}: 'Ꝼ'

LatinSmallLetterInsularF

\u{a77c}: 'ꝼ'

LatinCapitalLetterInsularG

\u{a77d}: 'Ᵹ'

LatinCapitalLetterTurnedInsularG

\u{a77e}: 'Ꝿ'

LatinSmallLetterTurnedInsularG

\u{a77f}: 'ꝿ'

LatinCapitalLetterTurnedL

\u{a780}: 'Ꞁ'

LatinSmallLetterTurnedL

\u{a781}: 'ꞁ'

LatinCapitalLetterInsularR

\u{a782}: 'Ꞃ'

LatinSmallLetterInsularR

\u{a783}: 'ꞃ'

LatinCapitalLetterInsularS

\u{a784}: 'Ꞅ'

LatinSmallLetterInsularS

\u{a785}: 'ꞅ'

LatinCapitalLetterInsularT

\u{a786}: 'Ꞇ'

LatinSmallLetterInsularT

\u{a787}: 'ꞇ'

ModifierLetterLowCircumflexAccent

\u{a788}: 'ꞈ'

ModifierLetterColon

\u{a789}: '꞉'

ModifierLetterShortEqualsSign

\u{a78a}: '꞊'

LatinCapitalLetterSaltillo

\u{a78b}: 'Ꞌ'

LatinSmallLetterSaltillo

\u{a78c}: 'ꞌ'

LatinCapitalLetterTurnedH

\u{a78d}: 'Ɥ'

LatinSmallLetterLWithRetroflexHookAndBelt

\u{a78e}: 'ꞎ'

LatinLetterSinologicalDot

\u{a78f}: 'ꞏ'

LatinCapitalLetterNWithDescender

\u{a790}: 'Ꞑ'

LatinSmallLetterNWithDescender

\u{a791}: 'ꞑ'

LatinCapitalLetterCWithBar

\u{a792}: 'Ꞓ'

LatinSmallLetterCWithBar

\u{a793}: 'ꞓ'

LatinSmallLetterCWithPalatalHook

\u{a794}: 'ꞔ'

LatinSmallLetterHWithPalatalHook

\u{a795}: 'ꞕ'

LatinCapitalLetterBWithFlourish

\u{a796}: 'Ꞗ'

LatinSmallLetterBWithFlourish

\u{a797}: 'ꞗ'

LatinCapitalLetterFWithStroke

\u{a798}: 'Ꞙ'

LatinSmallLetterFWithStroke

\u{a799}: 'ꞙ'

LatinCapitalLetterVolapukAe

\u{a79a}: 'Ꞛ'

LatinSmallLetterVolapukAe

\u{a79b}: 'ꞛ'

LatinCapitalLetterVolapukOe

\u{a79c}: 'Ꞝ'

LatinSmallLetterVolapukOe

\u{a79d}: 'ꞝ'

LatinCapitalLetterVolapukUe

\u{a79e}: 'Ꞟ'

LatinSmallLetterVolapukUe

\u{a79f}: 'ꞟ'

LatinCapitalLetterGWithObliqueStroke

\u{a7a0}: 'Ꞡ'

LatinSmallLetterGWithObliqueStroke

\u{a7a1}: 'ꞡ'

LatinCapitalLetterKWithObliqueStroke

\u{a7a2}: 'Ꞣ'

LatinSmallLetterKWithObliqueStroke

\u{a7a3}: 'ꞣ'

LatinCapitalLetterNWithObliqueStroke

\u{a7a4}: 'Ꞥ'

LatinSmallLetterNWithObliqueStroke

\u{a7a5}: 'ꞥ'

LatinCapitalLetterRWithObliqueStroke

\u{a7a6}: 'Ꞧ'

LatinSmallLetterRWithObliqueStroke

\u{a7a7}: 'ꞧ'

LatinCapitalLetterSWithObliqueStroke

\u{a7a8}: 'Ꞩ'

LatinSmallLetterSWithObliqueStroke

\u{a7a9}: 'ꞩ'

LatinCapitalLetterHWithHook

\u{a7aa}: 'Ɦ'

LatinCapitalLetterReversedOpenE

\u{a7ab}: 'Ɜ'

LatinCapitalLetterScriptG

\u{a7ac}: 'Ɡ'

LatinCapitalLetterLWithBelt

\u{a7ad}: 'Ɬ'

LatinCapitalLetterSmallCapitalI

\u{a7ae}: 'Ɪ'

LatinLetterSmallCapitalQ

\u{a7af}: 'ꞯ'

LatinCapitalLetterTurnedK

\u{a7b0}: 'Ʞ'

LatinCapitalLetterTurnedT

\u{a7b1}: 'Ʇ'

LatinCapitalLetterJWithCrossedDashTail

\u{a7b2}: 'Ʝ'

LatinCapitalLetterChi

\u{a7b3}: 'Ꭓ'

LatinCapitalLetterBeta

\u{a7b4}: 'Ꞵ'

LatinSmallLetterBeta

\u{a7b5}: 'ꞵ'

LatinCapitalLetterOmega

\u{a7b6}: 'Ꞷ'

LatinSmallLetterOmega

\u{a7b7}: 'ꞷ'

LatinCapitalLetterUWithStroke

\u{a7b8}: 'Ꞹ'

LatinSmallLetterUWithStroke

\u{a7b9}: 'ꞹ'

LatinCapitalLetterGlottalA

\u{a7ba}: 'Ꞻ'

LatinSmallLetterGlottalA

\u{a7bb}: 'ꞻ'

LatinCapitalLetterGlottalI

\u{a7bc}: 'Ꞽ'

LatinSmallLetterGlottalI

\u{a7bd}: 'ꞽ'

LatinCapitalLetterGlottalU

\u{a7be}: 'Ꞿ'

LatinSmallLetterGlottalU

\u{a7bf}: 'ꞿ'

LatinCapitalLetterAnglicanaW

\u{a7c2}: 'Ꟃ'

LatinSmallLetterAnglicanaW

\u{a7c3}: 'ꟃ'

LatinCapitalLetterCWithPalatalHook

\u{a7c4}: 'Ꞔ'

LatinCapitalLetterSWithHook

\u{a7c5}: 'Ʂ'

LatinCapitalLetterZWithPalatalHook

\u{a7c6}: 'Ᶎ'

LatinEpigraphicLetterSidewaysI

\u{a7f7}: 'ꟷ'

ModifierLetterCapitalHWithStroke

\u{a7f8}: 'ꟸ'

ModifierLetterSmallLigatureOe

\u{a7f9}: 'ꟹ'

LatinLetterSmallCapitalTurnedM

\u{a7fa}: 'ꟺ'

LatinEpigraphicLetterReversedF

\u{a7fb}: 'ꟻ'

LatinEpigraphicLetterReversedP

\u{a7fc}: 'ꟼ'

LatinEpigraphicLetterInvertedM

\u{a7fd}: 'ꟽ'

LatinEpigraphicLetterILonga

\u{a7fe}: 'ꟾ'

Methods

impl LatinExtendedD

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case