Enum unicode_types::linear_a::LinearA

pub enum LinearA {
    SignAb001,
    SignAb002,
    SignAb003,
    SignAb004,
    SignAb005,
    SignAb006,
    SignAb007,
    SignAb008,
    SignAb009,
    SignAb010,
    SignAb011,
    SignAb013,
    SignAb016,
    SignAb017,
    SignAb020,
    SignAb021,
    SignAb021f,
    SignAb021m,
    SignAb022,
    SignAb022f,
    SignAb022m,
    SignAb023,
    SignAb023m,
    SignAb024,
    SignAb026,
    SignAb027,
    SignAb028,
    SignA028b,
    SignAb029,
    SignAb030,
    SignAb031,
    SignAb034,
    SignAb037,
    SignAb038,
    SignAb039,
    SignAb040,
    SignAb041,
    SignAb044,
    SignAb045,
    SignAb046,
    SignAb047,
    SignAb048,
    SignAb049,
    SignAb050,
    SignAb051,
    SignAb053,
    SignAb054,
    SignAb055,
    SignAb056,
    SignAb057,
    SignAb058,
    SignAb059,
    SignAb060,
    SignAb061,
    SignAb065,
    SignAb066,
    SignAb067,
    SignAb069,
    SignAb070,
    SignAb073,
    SignAb074,
    SignAb076,
    SignAb077,
    SignAb078,
    SignAb079,
    SignAb080,
    SignAb081,
    SignAb082,
    SignAb085,
    SignAb086,
    SignAb087,
    SignA100Dash102,
    SignAb118,
    SignAb120,
    SignA120b,
    SignAb122,
    SignAb123,
    SignAb131a,
    SignAb131b,
    SignA131c,
    SignAb164,
    SignAb171,
    SignAb180,
    SignAb188,
    SignAb191,
    SignA301,
    SignA302,
    SignA303,
    SignA304,
    SignA305,
    SignA306,
    SignA307,
    SignA308,
    SignA309a,
    SignA309b,
    SignA309c,
    SignA310,
    SignA311,
    SignA312,
    SignA313a,
    SignA313b,
    SignA313c,
    SignA314,
    SignA315,
    SignA316,
    SignA317,
    SignA318,
    SignA319,
    SignA320,
    SignA321,
    SignA322,
    SignA323,
    SignA324,
    SignA325,
    SignA326,
    SignA327,
    SignA328,
    SignA329,
    SignA330,
    SignA331,
    SignA332,
    SignA333,
    SignA334,
    SignA335,
    SignA336,
    SignA337,
    SignA338,
    SignA339,
    SignA340,
    SignA341,
    SignA342,
    SignA343,
    SignA344,
    SignA345,
    SignA346,
    SignA347,
    SignA348,
    SignA349,
    SignA350,
    SignA351,
    SignA352,
    SignA353,
    SignA354,
    SignA355,
    SignA356,
    SignA357,
    SignA358,
    SignA359,
    SignA360,
    SignA361,
    SignA362,
    SignA363,
    SignA364,
    SignA365,
    SignA366,
    SignA367,
    SignA368,
    SignA369,
    SignA370,
    SignA371,
    SignA400DashVas,
    SignA401DashVas,
    SignA402DashVas,
    SignA403DashVas,
    SignA404DashVas,
    SignA405DashVas,
    SignA406DashVas,
    SignA407DashVas,
    SignA408DashVas,
    SignA409DashVas,
    SignA410DashVas,
    SignA411DashVas,
    SignA412DashVas,
    SignA413DashVas,
    SignA414DashVas,
    SignA415DashVas,
    SignA416DashVas,
    SignA417DashVas,
    SignA418DashVas,
    SignA501,
    SignA502,
    SignA503,
    SignA504,
    SignA505,
    SignA506,
    SignA508,
    SignA509,
    SignA510,
    SignA511,
    SignA512,
    SignA513,
    SignA515,
    SignA516,
    SignA520,
    SignA521,
    SignA523,
    SignA524,
    SignA525,
    SignA526,
    SignA527,
    SignA528,
    SignA529,
    SignA530,
    SignA531,
    SignA532,
    SignA534,
    SignA535,
    SignA536,
    SignA537,
    SignA538,
    SignA539,
    SignA540,
    SignA541,
    SignA542,
    SignA545,
    SignA547,
    SignA548,
    SignA549,
    SignA550,
    SignA551,
    SignA552,
    SignA553,
    SignA554,
    SignA555,
    SignA556,
    SignA557,
    SignA559,
    SignA563,
    SignA564,
    SignA565,
    SignA566,
    SignA568,
    SignA569,
    SignA570,
    SignA571,
    SignA572,
    SignA573,
    SignA574,
    SignA575,
    SignA576,
    SignA577,
    SignA578,
    SignA579,
    SignA580,
    SignA581,
    SignA582,
    SignA583,
    SignA584,
    SignA585,
    SignA586,
    SignA587,
    SignA588,
    SignA589,
    SignA591,
    SignA592,
    SignA594,
    SignA595,
    SignA596,
    SignA598,
    SignA600,
    SignA601,
    SignA602,
    SignA603,
    SignA604,
    SignA606,
    SignA608,
    SignA609,
    SignA610,
    SignA611,
    SignA612,
    SignA613,
    SignA614,
    SignA615,
    SignA616,
    SignA617,
    SignA618,
    SignA619,
    SignA620,
    SignA621,
    SignA622,
    SignA623,
    SignA624,
    SignA626,
    SignA627,
    SignA628,
    SignA629,
    SignA634,
    SignA637,
    SignA638,
    SignA640,
    SignA642,
    SignA643,
    SignA644,
    SignA645,
    SignA646,
    SignA648,
    SignA649,
    SignA651,
    SignA652,
    SignA653,
    SignA654,
    SignA655,
    SignA656,
    SignA657,
    SignA658,
    SignA659,
    SignA660,
    SignA661,
    SignA662,
    SignA663,
    SignA664,
    SignA701A,
    SignA702B,
    SignA703D,
    SignA704E,
    SignA705F,
    SignA706H,
    SignA707J,
    SignA708K,
    SignA709L,
    SignA709Dash2L2,
    SignA709Dash3L3,
    SignA709Dash4L4,
    SignA709Dash6L6,
    SignA710W,
    SignA711X,
    SignA712Y,
    SignA713Omega,
    SignA714Abb,
    SignA715Bb,
    SignA717Dd,
    SignA726Eyyy,
    SignA732Je,
    SignA800,
    SignA801,
    SignA802,
    SignA803,
    SignA804,
    SignA805,
    SignA806,
    SignA807,
}

An enum to represent all characters in the LinearA block.

Variants
SignAb001

\u{10600}: 'ğ˜€'

SignAb002

\u{10601}: 'ğ˜'

SignAb003

\u{10602}: 'ğ˜‚'

SignAb004

\u{10603}: 'ğ˜ƒ'

SignAb005

\u{10604}: 'ğ˜„'

SignAb006

\u{10605}: 'ğ˜…'

SignAb007

\u{10606}: 'ğ˜†'

SignAb008

\u{10607}: 'ğ˜‡'

SignAb009

\u{10608}: 'ğ˜ˆ'

SignAb010

\u{10609}: 'ğ˜‰'

SignAb011

\u{1060a}: 'ğ˜Š'

SignAb013

\u{1060b}: 'ğ˜‹'

SignAb016

\u{1060c}: 'ğ˜Œ'

SignAb017

\u{1060d}: 'ğ˜'

SignAb020

\u{1060e}: 'ğ˜'

SignAb021

\u{1060f}: 'ğ˜'

SignAb021f

\u{10610}: 'ğ˜'

SignAb021m

\u{10611}: 'ğ˜‘'

SignAb022

\u{10612}: 'ğ˜’'

SignAb022f

\u{10613}: 'ğ˜“'

SignAb022m

\u{10614}: 'ğ˜”'

SignAb023

\u{10615}: 'ğ˜•'

SignAb023m

\u{10616}: 'ğ˜–'

SignAb024

\u{10617}: 'ğ˜—'

SignAb026

\u{10618}: 'ğ˜˜'

SignAb027

\u{10619}: 'ğ˜™'

SignAb028

\u{1061a}: 'ğ˜š'

SignA028b

\u{1061b}: 'ğ˜›'

SignAb029

\u{1061c}: 'ğ˜œ'

SignAb030

\u{1061d}: 'ğ˜'

SignAb031

\u{1061e}: 'ğ˜'

SignAb034

\u{1061f}: 'ğ˜Ÿ'

SignAb037

\u{10620}: 'ğ˜ '

SignAb038

\u{10621}: 'ğ˜¡'

SignAb039

\u{10622}: 'ğ˜¢'

SignAb040

\u{10623}: 'ğ˜£'

SignAb041

\u{10624}: 'ğ˜¤'

SignAb044

\u{10625}: 'ğ˜¥'

SignAb045

\u{10626}: 'ğ˜¦'

SignAb046

\u{10627}: 'ğ˜'

SignAb047

\u{10628}: 'ğ˜¨'

SignAb048

\u{10629}: 'ğ˜©'

SignAb049

\u{1062a}: 'ğ˜ª'

SignAb050

\u{1062b}: 'ğ˜«'

SignAb051

\u{1062c}: 'ğ˜¬'

SignAb053

\u{1062d}: 'ğ˜­'

SignAb054

\u{1062e}: 'ğ˜®'

SignAb055

\u{1062f}: 'ğ˜¯'

SignAb056

\u{10630}: 'ğ˜°'

SignAb057

\u{10631}: 'ğ˜±'

SignAb058

\u{10632}: 'ğ˜²'

SignAb059

\u{10633}: 'ğ˜³'

SignAb060

\u{10634}: 'ğ˜´'

SignAb061

\u{10635}: 'ğ˜µ'

SignAb065

\u{10636}: 'ğ˜¶'

SignAb066

\u{10637}: 'ğ˜·'

SignAb067

\u{10638}: 'ğ˜¸'

SignAb069

\u{10639}: 'ğ˜¹'

SignAb070

\u{1063a}: 'ğ˜º'

SignAb073

\u{1063b}: 'ğ˜»'

SignAb074

\u{1063c}: 'ğ˜¼'

SignAb076

\u{1063d}: 'ğ˜½'

SignAb077

\u{1063e}: 'ğ˜¾'

SignAb078

\u{1063f}: 'ğ˜¿'

SignAb079

\u{10640}: 'ğ™€'

SignAb080

\u{10641}: 'ğ™'

SignAb081

\u{10642}: 'ğ™‚'

SignAb082

\u{10643}: 'ğ™ƒ'

SignAb085

\u{10644}: 'ğ™„'

SignAb086

\u{10645}: 'ğ™…'

SignAb087

\u{10646}: 'ğ™†'

SignA100Dash102

\u{10647}: 'ğ™‡'

SignAb118

\u{10648}: 'ğ™ˆ'

SignAb120

\u{10649}: 'ğ™‰'

SignA120b

\u{1064a}: 'ğ™Š'

SignAb122

\u{1064b}: 'ğ™‹'

SignAb123

\u{1064c}: 'ğ™Œ'

SignAb131a

\u{1064d}: 'ğ™'

SignAb131b

\u{1064e}: 'ğ™'

SignA131c

\u{1064f}: 'ğ™'

SignAb164

\u{10650}: 'ğ™'

SignAb171

\u{10651}: 'ğ™‘'

SignAb180

\u{10652}: 'ğ™’'

SignAb188

\u{10653}: 'ğ™“'

SignAb191

\u{10654}: 'ğ™”'

SignA301

\u{10655}: 'ğ™•'

SignA302

\u{10656}: 'ğ™–'

SignA303

\u{10657}: 'ğ™—'

SignA304

\u{10658}: 'ğ™˜'

SignA305

\u{10659}: 'ğ™™'

SignA306

\u{1065a}: 'ğ™š'

SignA307

\u{1065b}: 'ğ™›'

SignA308

\u{1065c}: 'ğ™œ'

SignA309a

\u{1065d}: 'ğ™'

SignA309b

\u{1065e}: 'ğ™'

SignA309c

\u{1065f}: 'ğ™Ÿ'

SignA310

\u{10660}: 'ğ™ '

SignA311

\u{10661}: 'ğ™¡'

SignA312

\u{10662}: 'ğ™¢'

SignA313a

\u{10663}: 'ğ™£'

SignA313b

\u{10664}: 'ğ™¤'

SignA313c

\u{10665}: 'ğ™¥'

SignA314

\u{10666}: 'ğ™¦'

SignA315

\u{10667}: 'ğ™'

SignA316

\u{10668}: 'ğ™¨'

SignA317

\u{10669}: 'ğ™©'

SignA318

\u{1066a}: 'ğ™ª'

SignA319

\u{1066b}: 'ğ™«'

SignA320

\u{1066c}: 'ğ™¬'

SignA321

\u{1066d}: 'ğ™­'

SignA322

\u{1066e}: 'ğ™®'

SignA323

\u{1066f}: 'ğ™¯'

SignA324

\u{10670}: 'ğ™°'

SignA325

\u{10671}: 'ğ™±'

SignA326

\u{10672}: 'ğ™²'

SignA327

\u{10673}: 'ğ™³'

SignA328

\u{10674}: 'ğ™´'

SignA329

\u{10675}: 'ğ™µ'

SignA330

\u{10676}: 'ğ™¶'

SignA331

\u{10677}: 'ğ™·'

SignA332

\u{10678}: 'ğ™¸'

SignA333

\u{10679}: 'ğ™¹'

SignA334

\u{1067a}: 'ğ™º'

SignA335

\u{1067b}: 'ğ™»'

SignA336

\u{1067c}: 'ğ™¼'

SignA337

\u{1067d}: 'ğ™½'

SignA338

\u{1067e}: 'ğ™¾'

SignA339

\u{1067f}: 'ğ™¿'

SignA340

\u{10680}: 'ğš€'

SignA341

\u{10681}: 'ğš'

SignA342

\u{10682}: 'ğš‚'

SignA343

\u{10683}: 'ğšƒ'

SignA344

\u{10684}: 'ğš„'

SignA345

\u{10685}: 'ğš…'

SignA346

\u{10686}: 'ğš†'

SignA347

\u{10687}: 'ğš‡'

SignA348

\u{10688}: 'ğšˆ'

SignA349

\u{10689}: 'ğš‰'

SignA350

\u{1068a}: 'ğšŠ'

SignA351

\u{1068b}: 'ğš‹'

SignA352

\u{1068c}: 'ğšŒ'

SignA353

\u{1068d}: 'ğš'

SignA354

\u{1068e}: 'ğš'

SignA355

\u{1068f}: 'ğš'

SignA356

\u{10690}: 'ğš'

SignA357

\u{10691}: 'ğš‘'

SignA358

\u{10692}: 'ğš’'

SignA359

\u{10693}: 'ğš“'

SignA360

\u{10694}: 'ğš”'

SignA361

\u{10695}: 'ğš•'

SignA362

\u{10696}: 'ğš–'

SignA363

\u{10697}: 'ğš—'

SignA364

\u{10698}: 'ğš˜'

SignA365

\u{10699}: 'ğš™'

SignA366

\u{1069a}: 'ğšš'

SignA367

\u{1069b}: 'ğš›'

SignA368

\u{1069c}: 'ğšœ'

SignA369

\u{1069d}: 'ğš'

SignA370

\u{1069e}: 'ğš'

SignA371

\u{1069f}: 'ğšŸ'

SignA400DashVas

\u{106a0}: 'ğš '

SignA401DashVas

\u{106a1}: 'ğš¡'

SignA402DashVas

\u{106a2}: 'ğš¢'

SignA403DashVas

\u{106a3}: 'ğš£'

SignA404DashVas

\u{106a4}: 'ğš¤'

SignA405DashVas

\u{106a5}: 'ğš¥'

SignA406DashVas

\u{106a6}: 'ğš¦'

SignA407DashVas

\u{106a7}: 'ğš'

SignA408DashVas

\u{106a8}: 'ğš¨'

SignA409DashVas

\u{106a9}: 'ğš©'

SignA410DashVas

\u{106aa}: 'ğšª'

SignA411DashVas

\u{106ab}: 'ğš«'

SignA412DashVas

\u{106ac}: 'ğš¬'

SignA413DashVas

\u{106ad}: 'ğš­'

SignA414DashVas

\u{106ae}: 'ğš®'

SignA415DashVas

\u{106af}: 'ğš¯'

SignA416DashVas

\u{106b0}: 'ğš°'

SignA417DashVas

\u{106b1}: 'ğš±'

SignA418DashVas

\u{106b2}: 'ğš²'

SignA501

\u{106b3}: 'ğš³'

SignA502

\u{106b4}: 'ğš´'

SignA503

\u{106b5}: 'ğšµ'

SignA504

\u{106b6}: 'ğš¶'

SignA505

\u{106b7}: 'ğš·'

SignA506

\u{106b8}: 'ğš¸'

SignA508

\u{106b9}: 'ğš¹'

SignA509

\u{106ba}: 'ğšº'

SignA510

\u{106bb}: 'ğš»'

SignA511

\u{106bc}: 'ğš¼'

SignA512

\u{106bd}: 'ğš½'

SignA513

\u{106be}: 'ğš¾'

SignA515

\u{106bf}: 'ğš¿'

SignA516

\u{106c0}: 'ğ›€'

SignA520

\u{106c1}: 'ğ›'

SignA521

\u{106c2}: 'ğ›‚'

SignA523

\u{106c3}: 'ğ›ƒ'

SignA524

\u{106c4}: 'ğ›„'

SignA525

\u{106c5}: 'ğ›…'

SignA526

\u{106c6}: 'ğ›†'

SignA527

\u{106c7}: 'ğ›‡'

SignA528

\u{106c8}: 'ğ›ˆ'

SignA529

\u{106c9}: 'ğ›‰'

SignA530

\u{106ca}: 'ğ›Š'

SignA531

\u{106cb}: 'ğ›‹'

SignA532

\u{106cc}: 'ğ›Œ'

SignA534

\u{106cd}: 'ğ›'

SignA535

\u{106ce}: 'ğ›'

SignA536

\u{106cf}: 'ğ›'

SignA537

\u{106d0}: 'ğ›'

SignA538

\u{106d1}: 'ğ›‘'

SignA539

\u{106d2}: 'ğ›’'

SignA540

\u{106d3}: 'ğ›“'

SignA541

\u{106d4}: 'ğ›”'

SignA542

\u{106d5}: 'ğ›•'

SignA545

\u{106d6}: 'ğ›–'

SignA547

\u{106d7}: 'ğ›—'

SignA548

\u{106d8}: 'ğ›˜'

SignA549

\u{106d9}: 'ğ›™'

SignA550

\u{106da}: 'ğ›š'

SignA551

\u{106db}: 'ğ››'

SignA552

\u{106dc}: 'ğ›œ'

SignA553

\u{106dd}: 'ğ›'

SignA554

\u{106de}: 'ğ›'

SignA555

\u{106df}: 'ğ›Ÿ'

SignA556

\u{106e0}: 'ğ› '

SignA557

\u{106e1}: 'ğ›¡'

SignA559

\u{106e2}: 'ğ›¢'

SignA563

\u{106e3}: 'ğ›£'

SignA564

\u{106e4}: 'ğ›¤'

SignA565

\u{106e5}: 'ğ›¥'

SignA566

\u{106e6}: 'ğ›¦'

SignA568

\u{106e7}: 'ğ›'

SignA569

\u{106e8}: 'ğ›¨'

SignA570

\u{106e9}: 'ğ›©'

SignA571

\u{106ea}: 'ğ›ª'

SignA572

\u{106eb}: 'ğ›«'

SignA573

\u{106ec}: 'ğ›¬'

SignA574

\u{106ed}: 'ğ›­'

SignA575

\u{106ee}: 'ğ›®'

SignA576

\u{106ef}: 'ğ›¯'

SignA577

\u{106f0}: 'ğ›°'

SignA578

\u{106f1}: 'ğ›±'

SignA579

\u{106f2}: 'ğ›²'

SignA580

\u{106f3}: 'ğ›³'

SignA581

\u{106f4}: 'ğ›´'

SignA582

\u{106f5}: 'ğ›µ'

SignA583

\u{106f6}: 'ğ›¶'

SignA584

\u{106f7}: 'ğ›·'

SignA585

\u{106f8}: 'ğ›¸'

SignA586

\u{106f9}: 'ğ›¹'

SignA587

\u{106fa}: 'ğ›º'

SignA588

\u{106fb}: 'ğ›»'

SignA589

\u{106fc}: 'ğ›¼'

SignA591

\u{106fd}: 'ğ›½'

SignA592

\u{106fe}: 'ğ›¾'

SignA594

\u{106ff}: 'ğ›¿'

SignA595

\u{10700}: 'ğœ€'

SignA596

\u{10701}: 'ğœ'

SignA598

\u{10702}: 'ğœ‚'

SignA600

\u{10703}: 'ğœƒ'

SignA601

\u{10704}: 'ğœ„'

SignA602

\u{10705}: 'ğœ…'

SignA603

\u{10706}: 'ğœ†'

SignA604

\u{10707}: 'ğœ‡'

SignA606

\u{10708}: 'ğœˆ'

SignA608

\u{10709}: 'ğœ‰'

SignA609

\u{1070a}: 'ğœŠ'

SignA610

\u{1070b}: 'ğœ‹'

SignA611

\u{1070c}: 'ğœŒ'

SignA612

\u{1070d}: 'ğœ'

SignA613

\u{1070e}: 'ğœ'

SignA614

\u{1070f}: 'ğœ'

SignA615

\u{10710}: 'ğœ'

SignA616

\u{10711}: 'ğœ‘'

SignA617

\u{10712}: 'ğœ’'

SignA618

\u{10713}: 'ğœ“'

SignA619

\u{10714}: 'ğœ”'

SignA620

\u{10715}: 'ğœ•'

SignA621

\u{10716}: 'ğœ–'

SignA622

\u{10717}: 'ğœ—'

SignA623

\u{10718}: 'ğœ˜'

SignA624

\u{10719}: 'ğœ™'

SignA626

\u{1071a}: 'ğœš'

SignA627

\u{1071b}: 'ğœ›'

SignA628

\u{1071c}: 'ğœœ'

SignA629

\u{1071d}: 'ğœ'

SignA634

\u{1071e}: 'ğœ'

SignA637

\u{1071f}: 'ğœŸ'

SignA638

\u{10720}: 'ğœ '

SignA640

\u{10721}: 'ğœ¡'

SignA642

\u{10722}: 'ğœ¢'

SignA643

\u{10723}: 'ğœ£'

SignA644

\u{10724}: 'ğœ¤'

SignA645

\u{10725}: 'ğœ¥'

SignA646

\u{10726}: 'ğœ¦'

SignA648

\u{10727}: 'ğœ'

SignA649

\u{10728}: 'ğœ¨'

SignA651

\u{10729}: 'ğœ©'

SignA652

\u{1072a}: 'ğœª'

SignA653

\u{1072b}: 'ğœ«'

SignA654

\u{1072c}: 'ğœ¬'

SignA655

\u{1072d}: 'ğœ­'

SignA656

\u{1072e}: 'ğœ®'

SignA657

\u{1072f}: 'ğœ¯'

SignA658

\u{10730}: 'ğœ°'

SignA659

\u{10731}: 'ğœ±'

SignA660

\u{10732}: 'ğœ²'

SignA661

\u{10733}: 'ğœ³'

SignA662

\u{10734}: 'ğœ´'

SignA663

\u{10735}: 'ğœµ'

SignA664

\u{10736}: 'ğœ¶'

SignA701A

\u{10740}: 'ğ€'

SignA702B

\u{10741}: 'ğ'

SignA703D

\u{10742}: 'ğ‚'

SignA704E

\u{10743}: 'ğƒ'

SignA705F

\u{10744}: 'ğ„'

SignA706H

\u{10745}: 'ğ…'

SignA707J

\u{10746}: 'ğ†'

SignA708K

\u{10747}: 'ğ‡'

SignA709L

\u{10748}: 'ğˆ'

SignA709Dash2L2

\u{10749}: 'ğ‰'

SignA709Dash3L3

\u{1074a}: 'ğŠ'

SignA709Dash4L4

\u{1074b}: 'ğ‹'

SignA709Dash6L6

\u{1074c}: 'ğŒ'

SignA710W

\u{1074d}: 'ğ'

SignA711X

\u{1074e}: 'ğ'

SignA712Y

\u{1074f}: 'ğ'

SignA713Omega

\u{10750}: 'ğ'

SignA714Abb

\u{10751}: 'ğ‘'

SignA715Bb

\u{10752}: 'ğ’'

SignA717Dd

\u{10753}: 'ğ“'

SignA726Eyyy

\u{10754}: 'ğ”'

SignA732Je

\u{10755}: 'ğ•'

SignA800

\u{10760}: 'ğ '

SignA801

\u{10761}: 'ğ¡'

SignA802

\u{10762}: 'ğ¢'

SignA803

\u{10763}: 'ğ£'

SignA804

\u{10764}: 'ğ¤'

SignA805

\u{10765}: 'ğ¥'

SignA806

\u{10766}: 'ğ¦'

SignA807

\u{10767}: 'ğ'

Methods

impl LinearA

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case