Enum unicode_types::linear_b_ideograms::LinearBIdeograms

pub enum LinearBIdeograms {
    LinearBIdeogramB100Man,
    LinearBIdeogramB102Woman,
    LinearBIdeogramB104Deer,
    LinearBIdeogramB105Equid,
    LinearBIdeogramB105fMare,
    LinearBIdeogramB105mStallion,
    LinearBIdeogramB106fEwe,
    LinearBIdeogramB106mRam,
    LinearBIdeogramB107fSheDashGoat,
    LinearBIdeogramB107mHeDashGoat,
    LinearBIdeogramB108fSow,
    LinearBIdeogramB108mBoar,
    LinearBIdeogramB109fCow,
    LinearBIdeogramB109mBull,
    LinearBIdeogramB120Wheat,
    LinearBIdeogramB121Barley,
    LinearBIdeogramB122Olive,
    LinearBIdeogramB123Spice,
    LinearBIdeogramB125Cyperus,
    LinearBMonogramB127Kapo,
    LinearBMonogramB128Kanako,
    LinearBIdeogramB130Oil,
    LinearBIdeogramB131Wine,
    LinearBIdeogramB132,
    LinearBMonogramB133Arepa,
    LinearBMonogramB135Meri,
    LinearBIdeogramB140Bronze,
    LinearBIdeogramB141Gold,
    LinearBIdeogramB142,
    LinearBIdeogramB145Wool,
    LinearBIdeogramB146,
    LinearBIdeogramB150,
    LinearBIdeogramB151Horn,
    LinearBIdeogramB152,
    LinearBIdeogramB153,
    LinearBIdeogramB154,
    LinearBMonogramB156Turo2,
    LinearBIdeogramB157,
    LinearBIdeogramB158,
    LinearBIdeogramB159Cloth,
    LinearBIdeogramB160,
    LinearBIdeogramB161,
    LinearBIdeogramB162Garment,
    LinearBIdeogramB163Armour,
    LinearBIdeogramB164,
    LinearBIdeogramB165,
    LinearBIdeogramB166,
    LinearBIdeogramB167,
    LinearBIdeogramB168,
    LinearBIdeogramB169,
    LinearBIdeogramB170,
    LinearBIdeogramB171,
    LinearBIdeogramB172,
    LinearBIdeogramB173Month,
    LinearBIdeogramB174,
    LinearBIdeogramB176Tree,
    LinearBIdeogramB177,
    LinearBIdeogramB178,
    LinearBIdeogramB179,
    LinearBIdeogramB180,
    LinearBIdeogramB181,
    LinearBIdeogramB182,
    LinearBIdeogramB183,
    LinearBIdeogramB184,
    LinearBIdeogramB185,
    LinearBIdeogramB189,
    LinearBIdeogramB190,
    LinearBIdeogramB191Helmet,
    LinearBIdeogramB220Footstool,
    LinearBIdeogramB225Bathtub,
    LinearBIdeogramB230Spear,
    LinearBIdeogramB231Arrow,
    LinearBIdeogramB232,
    LinearBIdeogramB233Sword,
    LinearBIdeogramB234,
    LinearBIdeogramB236,
    LinearBIdeogramB240WheeledChariot,
    LinearBIdeogramB241Chariot,
    LinearBIdeogramB242ChariotFrame,
    LinearBIdeogramB243Wheel,
    LinearBIdeogramB245,
    LinearBIdeogramB246,
    LinearBMonogramB247Dipte,
    LinearBIdeogramB248,
    LinearBIdeogramB249,
    LinearBIdeogramB251,
    LinearBIdeogramB252,
    LinearBIdeogramB253,
    LinearBIdeogramB254Dart,
    LinearBIdeogramB255,
    LinearBIdeogramB256,
    LinearBIdeogramB257,
    LinearBIdeogramB258,
    LinearBIdeogramB259,
    LinearBIdeogramVesselB155,
    LinearBIdeogramVesselB200,
    LinearBIdeogramVesselB201,
    LinearBIdeogramVesselB202,
    LinearBIdeogramVesselB203,
    LinearBIdeogramVesselB204,
    LinearBIdeogramVesselB205,
    LinearBIdeogramVesselB206,
    LinearBIdeogramVesselB207,
    LinearBIdeogramVesselB208,
    LinearBIdeogramVesselB209,
    LinearBIdeogramVesselB210,
    LinearBIdeogramVesselB211,
    LinearBIdeogramVesselB212,
    LinearBIdeogramVesselB213,
    LinearBIdeogramVesselB214,
    LinearBIdeogramVesselB215,
    LinearBIdeogramVesselB216,
    LinearBIdeogramVesselB217,
    LinearBIdeogramVesselB218,
    LinearBIdeogramVesselB219,
    LinearBIdeogramVesselB221,
    LinearBIdeogramVesselB222,
    LinearBIdeogramVesselB226,
    LinearBIdeogramVesselB227,
    LinearBIdeogramVesselB228,
    LinearBIdeogramVesselB229,
    LinearBIdeogramVesselB250,
    LinearBIdeogramVesselB305,
}

An enum to represent all characters in the LinearBIdeograms block.

Variants
LinearBIdeogramB100Man

\u{10080}: 'ğ‚€'

LinearBIdeogramB102Woman

\u{10081}: 'ğ‚'

LinearBIdeogramB104Deer

\u{10082}: 'ğ‚‚'

LinearBIdeogramB105Equid

\u{10083}: 'ğ‚ƒ'

LinearBIdeogramB105fMare

\u{10084}: 'ğ‚„'

LinearBIdeogramB105mStallion

\u{10085}: 'ğ‚…'

LinearBIdeogramB106fEwe

\u{10086}: 'ğ‚†'

LinearBIdeogramB106mRam

\u{10087}: 'ğ‚‡'

LinearBIdeogramB107fSheDashGoat

\u{10088}: 'ğ‚ˆ'

LinearBIdeogramB107mHeDashGoat

\u{10089}: 'ğ‚‰'

LinearBIdeogramB108fSow

\u{1008a}: 'ğ‚Š'

LinearBIdeogramB108mBoar

\u{1008b}: 'ğ‚‹'

LinearBIdeogramB109fCow

\u{1008c}: 'ğ‚Œ'

LinearBIdeogramB109mBull

\u{1008d}: 'ğ‚'

LinearBIdeogramB120Wheat

\u{1008e}: 'ğ‚'

LinearBIdeogramB121Barley

\u{1008f}: 'ğ‚'

LinearBIdeogramB122Olive

\u{10090}: 'ğ‚'

LinearBIdeogramB123Spice

\u{10091}: 'ğ‚‘'

LinearBIdeogramB125Cyperus

\u{10092}: 'ğ‚’'

LinearBMonogramB127Kapo

\u{10093}: 'ğ‚“'

LinearBMonogramB128Kanako

\u{10094}: 'ğ‚”'

LinearBIdeogramB130Oil

\u{10095}: 'ğ‚•'

LinearBIdeogramB131Wine

\u{10096}: 'ğ‚–'

LinearBIdeogramB132

\u{10097}: 'ğ‚—'

LinearBMonogramB133Arepa

\u{10098}: 'ğ‚˜'

LinearBMonogramB135Meri

\u{10099}: 'ğ‚™'

LinearBIdeogramB140Bronze

\u{1009a}: 'ğ‚š'

LinearBIdeogramB141Gold

\u{1009b}: 'ğ‚›'

LinearBIdeogramB142

\u{1009c}: 'ğ‚œ'

LinearBIdeogramB145Wool

\u{1009d}: 'ğ‚'

LinearBIdeogramB146

\u{1009e}: 'ğ‚'

LinearBIdeogramB150

\u{1009f}: 'ğ‚Ÿ'

LinearBIdeogramB151Horn

\u{100a0}: 'ğ‚ '

LinearBIdeogramB152

\u{100a1}: 'ğ‚¡'

LinearBIdeogramB153

\u{100a2}: 'ğ‚¢'

LinearBIdeogramB154

\u{100a3}: 'ğ‚£'

LinearBMonogramB156Turo2

\u{100a4}: 'ğ‚¤'

LinearBIdeogramB157

\u{100a5}: 'ğ‚¥'

LinearBIdeogramB158

\u{100a6}: 'ğ‚¦'

LinearBIdeogramB159Cloth

\u{100a7}: 'ğ‚'

LinearBIdeogramB160

\u{100a8}: 'ğ‚¨'

LinearBIdeogramB161

\u{100a9}: 'ğ‚©'

LinearBIdeogramB162Garment

\u{100aa}: 'ğ‚ª'

LinearBIdeogramB163Armour

\u{100ab}: 'ğ‚«'

LinearBIdeogramB164

\u{100ac}: 'ğ‚¬'

LinearBIdeogramB165

\u{100ad}: 'ğ‚­'

LinearBIdeogramB166

\u{100ae}: 'ğ‚®'

LinearBIdeogramB167

\u{100af}: 'ğ‚¯'

LinearBIdeogramB168

\u{100b0}: 'ğ‚°'

LinearBIdeogramB169

\u{100b1}: 'ğ‚±'

LinearBIdeogramB170

\u{100b2}: 'ğ‚²'

LinearBIdeogramB171

\u{100b3}: 'ğ‚³'

LinearBIdeogramB172

\u{100b4}: 'ğ‚´'

LinearBIdeogramB173Month

\u{100b5}: 'ğ‚µ'

LinearBIdeogramB174

\u{100b6}: 'ğ‚¶'

LinearBIdeogramB176Tree

\u{100b7}: 'ğ‚·'

LinearBIdeogramB177

\u{100b8}: 'ğ‚¸'

LinearBIdeogramB178

\u{100b9}: 'ğ‚¹'

LinearBIdeogramB179

\u{100ba}: 'ğ‚º'

LinearBIdeogramB180

\u{100bb}: 'ğ‚»'

LinearBIdeogramB181

\u{100bc}: 'ğ‚¼'

LinearBIdeogramB182

\u{100bd}: 'ğ‚½'

LinearBIdeogramB183

\u{100be}: 'ğ‚¾'

LinearBIdeogramB184

\u{100bf}: 'ğ‚¿'

LinearBIdeogramB185

\u{100c0}: 'ğƒ€'

LinearBIdeogramB189

\u{100c1}: 'ğƒ'

LinearBIdeogramB190

\u{100c2}: 'ğƒ‚'

LinearBIdeogramB191Helmet

\u{100c3}: 'ğƒƒ'

LinearBIdeogramB220Footstool

\u{100c4}: 'ğƒ„'

LinearBIdeogramB225Bathtub

\u{100c5}: 'ğƒ…'

LinearBIdeogramB230Spear

\u{100c6}: 'ğƒ†'

LinearBIdeogramB231Arrow

\u{100c7}: 'ğƒ‡'

LinearBIdeogramB232

\u{100c8}: 'ğƒˆ'

LinearBIdeogramB233Sword

\u{100c9}: 'ğƒ‰'

LinearBIdeogramB234

\u{100ca}: 'ğƒŠ'

LinearBIdeogramB236

\u{100cb}: 'ğƒ‹'

LinearBIdeogramB240WheeledChariot

\u{100cc}: 'ğƒŒ'

LinearBIdeogramB241Chariot

\u{100cd}: 'ğƒ'

LinearBIdeogramB242ChariotFrame

\u{100ce}: 'ğƒ'

LinearBIdeogramB243Wheel

\u{100cf}: 'ğƒ'

LinearBIdeogramB245

\u{100d0}: 'ğƒ'

LinearBIdeogramB246

\u{100d1}: 'ğƒ‘'

LinearBMonogramB247Dipte

\u{100d2}: 'ğƒ’'

LinearBIdeogramB248

\u{100d3}: 'ğƒ“'

LinearBIdeogramB249

\u{100d4}: 'ğƒ”'

LinearBIdeogramB251

\u{100d5}: 'ğƒ•'

LinearBIdeogramB252

\u{100d6}: 'ğƒ–'

LinearBIdeogramB253

\u{100d7}: 'ğƒ—'

LinearBIdeogramB254Dart

\u{100d8}: 'ğƒ˜'

LinearBIdeogramB255

\u{100d9}: 'ğƒ™'

LinearBIdeogramB256

\u{100da}: 'ğƒš'

LinearBIdeogramB257

\u{100db}: 'ğƒ›'

LinearBIdeogramB258

\u{100dc}: 'ğƒœ'

LinearBIdeogramB259

\u{100dd}: 'ğƒ'

LinearBIdeogramVesselB155

\u{100de}: 'ğƒ'

LinearBIdeogramVesselB200

\u{100df}: 'ğƒŸ'

LinearBIdeogramVesselB201

\u{100e0}: 'ğƒ '

LinearBIdeogramVesselB202

\u{100e1}: 'ğƒ¡'

LinearBIdeogramVesselB203

\u{100e2}: 'ğƒ¢'

LinearBIdeogramVesselB204

\u{100e3}: 'ğƒ£'

LinearBIdeogramVesselB205

\u{100e4}: 'ğƒ¤'

LinearBIdeogramVesselB206

\u{100e5}: 'ğƒ¥'

LinearBIdeogramVesselB207

\u{100e6}: 'ğƒ¦'

LinearBIdeogramVesselB208

\u{100e7}: 'ğƒ'

LinearBIdeogramVesselB209

\u{100e8}: 'ğƒ¨'

LinearBIdeogramVesselB210

\u{100e9}: 'ğƒ©'

LinearBIdeogramVesselB211

\u{100ea}: 'ğƒª'

LinearBIdeogramVesselB212

\u{100eb}: 'ğƒ«'

LinearBIdeogramVesselB213

\u{100ec}: 'ğƒ¬'

LinearBIdeogramVesselB214

\u{100ed}: 'ğƒ­'

LinearBIdeogramVesselB215

\u{100ee}: 'ğƒ®'

LinearBIdeogramVesselB216

\u{100ef}: 'ğƒ¯'

LinearBIdeogramVesselB217

\u{100f0}: 'ğƒ°'

LinearBIdeogramVesselB218

\u{100f1}: 'ğƒ±'

LinearBIdeogramVesselB219

\u{100f2}: 'ğƒ²'

LinearBIdeogramVesselB221

\u{100f3}: 'ğƒ³'

LinearBIdeogramVesselB222

\u{100f4}: 'ğƒ´'

LinearBIdeogramVesselB226

\u{100f5}: 'ğƒµ'

LinearBIdeogramVesselB227

\u{100f6}: 'ğƒ¶'

LinearBIdeogramVesselB228

\u{100f7}: 'ğƒ·'

LinearBIdeogramVesselB229

\u{100f8}: 'ğƒ¸'

LinearBIdeogramVesselB250

\u{100f9}: 'ğƒ¹'

LinearBIdeogramVesselB305

\u{100fa}: 'ğƒº'

Methods

impl LinearBIdeograms

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case