Enum unicode_types::linear_b_syllabary::LinearBSyllabary

pub enum LinearBSyllabary {
    LinearBSyllableB008A,
    LinearBSyllableB038E,
    LinearBSyllableB028I,
    LinearBSyllableB061O,
    LinearBSyllableB010U,
    LinearBSyllableB001Da,
    LinearBSyllableB045De,
    LinearBSyllableB007Di,
    LinearBSyllableB014Do,
    LinearBSyllableB051Du,
    LinearBSyllableB057Ja,
    LinearBSyllableB046Je,
    LinearBSyllableB036Jo,
    LinearBSyllableB065Ju,
    LinearBSyllableB077Ka,
    LinearBSyllableB044Ke,
    LinearBSyllableB067Ki,
    LinearBSyllableB070Ko,
    LinearBSyllableB081Ku,
    LinearBSyllableB080Ma,
    LinearBSyllableB013Me,
    LinearBSyllableB073Mi,
    LinearBSyllableB015Mo,
    LinearBSyllableB023Mu,
    LinearBSyllableB006Na,
    LinearBSyllableB024Ne,
    LinearBSyllableB030Ni,
    LinearBSyllableB052No,
    LinearBSyllableB055Nu,
    LinearBSyllableB003Pa,
    LinearBSyllableB072Pe,
    LinearBSyllableB039Pi,
    LinearBSyllableB011Po,
    LinearBSyllableB050Pu,
    LinearBSyllableB016Qa,
    LinearBSyllableB078Qe,
    LinearBSyllableB021Qi,
    LinearBSyllableB032Qo,
    LinearBSyllableB060Ra,
    LinearBSyllableB027Re,
    LinearBSyllableB053Ri,
    LinearBSyllableB002Ro,
    LinearBSyllableB026Ru,
    LinearBSyllableB031Sa,
    LinearBSyllableB009Se,
    LinearBSyllableB041Si,
    LinearBSyllableB012So,
    LinearBSyllableB058Su,
    LinearBSyllableB059Ta,
    LinearBSyllableB004Te,
    LinearBSyllableB037Ti,
    LinearBSyllableB005To,
    LinearBSyllableB069Tu,
    LinearBSyllableB054Wa,
    LinearBSyllableB075We,
    LinearBSyllableB040Wi,
    LinearBSyllableB042Wo,
    LinearBSyllableB017Za,
    LinearBSyllableB074Ze,
    LinearBSyllableB020Zo,
    LinearBSyllableB025A2,
    LinearBSyllableB043A3,
    LinearBSyllableB085Au,
    LinearBSyllableB071Dwe,
    LinearBSyllableB090Dwo,
    LinearBSyllableB048Nwa,
    LinearBSyllableB029Pu2,
    LinearBSyllableB062Pte,
    LinearBSyllableB076Ra2,
    LinearBSyllableB033Ra3,
    LinearBSyllableB068Ro2,
    LinearBSyllableB066Ta2,
    LinearBSyllableB087Twe,
    LinearBSyllableB091Two,
    LinearBSymbolB018,
    LinearBSymbolB019,
    LinearBSymbolB022,
    LinearBSymbolB034,
    LinearBSymbolB047,
    LinearBSymbolB049,
    LinearBSymbolB056,
    LinearBSymbolB063,
    LinearBSymbolB064,
    LinearBSymbolB079,
    LinearBSymbolB082,
    LinearBSymbolB083,
    LinearBSymbolB086,
    LinearBSymbolB089,
}

An enum to represent all characters in the LinearBSyllabary block.

Variants
LinearBSyllableB008A

\u{10000}: 'ğ€€'

LinearBSyllableB038E

\u{10001}: 'ğ€'

LinearBSyllableB028I

\u{10002}: 'ğ€‚'

LinearBSyllableB061O

\u{10003}: 'ğ€ƒ'

LinearBSyllableB010U

\u{10004}: 'ğ€„'

LinearBSyllableB001Da

\u{10005}: 'ğ€…'

LinearBSyllableB045De

\u{10006}: 'ğ€†'

LinearBSyllableB007Di

\u{10007}: 'ğ€‡'

LinearBSyllableB014Do

\u{10008}: 'ğ€ˆ'

LinearBSyllableB051Du

\u{10009}: 'ğ€‰'

LinearBSyllableB057Ja

\u{1000a}: 'ğ€Š'

LinearBSyllableB046Je

\u{1000b}: 'ğ€‹'

LinearBSyllableB036Jo

\u{1000d}: 'ğ€'

LinearBSyllableB065Ju

\u{1000e}: 'ğ€'

LinearBSyllableB077Ka

\u{1000f}: 'ğ€'

LinearBSyllableB044Ke

\u{10010}: 'ğ€'

LinearBSyllableB067Ki

\u{10011}: 'ğ€‘'

LinearBSyllableB070Ko

\u{10012}: 'ğ€’'

LinearBSyllableB081Ku

\u{10013}: 'ğ€“'

LinearBSyllableB080Ma

\u{10014}: 'ğ€”'

LinearBSyllableB013Me

\u{10015}: 'ğ€•'

LinearBSyllableB073Mi

\u{10016}: 'ğ€–'

LinearBSyllableB015Mo

\u{10017}: 'ğ€—'

LinearBSyllableB023Mu

\u{10018}: 'ğ€˜'

LinearBSyllableB006Na

\u{10019}: 'ğ€™'

LinearBSyllableB024Ne

\u{1001a}: 'ğ€š'

LinearBSyllableB030Ni

\u{1001b}: 'ğ€›'

LinearBSyllableB052No

\u{1001c}: 'ğ€œ'

LinearBSyllableB055Nu

\u{1001d}: 'ğ€'

LinearBSyllableB003Pa

\u{1001e}: 'ğ€'

LinearBSyllableB072Pe

\u{1001f}: 'ğ€Ÿ'

LinearBSyllableB039Pi

\u{10020}: 'ğ€ '

LinearBSyllableB011Po

\u{10021}: 'ğ€¡'

LinearBSyllableB050Pu

\u{10022}: 'ğ€¢'

LinearBSyllableB016Qa

\u{10023}: 'ğ€£'

LinearBSyllableB078Qe

\u{10024}: 'ğ€¤'

LinearBSyllableB021Qi

\u{10025}: 'ğ€¥'

LinearBSyllableB032Qo

\u{10026}: 'ğ€¦'

LinearBSyllableB060Ra

\u{10028}: 'ğ€¨'

LinearBSyllableB027Re

\u{10029}: 'ğ€©'

LinearBSyllableB053Ri

\u{1002a}: 'ğ€ª'

LinearBSyllableB002Ro

\u{1002b}: 'ğ€«'

LinearBSyllableB026Ru

\u{1002c}: 'ğ€¬'

LinearBSyllableB031Sa

\u{1002d}: 'ğ€­'

LinearBSyllableB009Se

\u{1002e}: 'ğ€®'

LinearBSyllableB041Si

\u{1002f}: 'ğ€¯'

LinearBSyllableB012So

\u{10030}: 'ğ€°'

LinearBSyllableB058Su

\u{10031}: 'ğ€±'

LinearBSyllableB059Ta

\u{10032}: 'ğ€²'

LinearBSyllableB004Te

\u{10033}: 'ğ€³'

LinearBSyllableB037Ti

\u{10034}: 'ğ€´'

LinearBSyllableB005To

\u{10035}: 'ğ€µ'

LinearBSyllableB069Tu

\u{10036}: 'ğ€¶'

LinearBSyllableB054Wa

\u{10037}: 'ğ€·'

LinearBSyllableB075We

\u{10038}: 'ğ€¸'

LinearBSyllableB040Wi

\u{10039}: 'ğ€¹'

LinearBSyllableB042Wo

\u{1003a}: 'ğ€º'

LinearBSyllableB017Za

\u{1003c}: 'ğ€¼'

LinearBSyllableB074Ze

\u{1003d}: 'ğ€½'

LinearBSyllableB020Zo

\u{1003f}: 'ğ€¿'

LinearBSyllableB025A2

\u{10040}: 'ğ€'

LinearBSyllableB043A3

\u{10041}: 'ğ'

LinearBSyllableB085Au

\u{10042}: 'ğ‚'

LinearBSyllableB071Dwe

\u{10043}: 'ğƒ'

LinearBSyllableB090Dwo

\u{10044}: 'ğ„'

LinearBSyllableB048Nwa

\u{10045}: 'ğ…'

LinearBSyllableB029Pu2

\u{10046}: 'ğ†'

LinearBSyllableB062Pte

\u{10047}: 'ğ‡'

LinearBSyllableB076Ra2

\u{10048}: 'ğˆ'

LinearBSyllableB033Ra3

\u{10049}: 'ğ‰'

LinearBSyllableB068Ro2

\u{1004a}: 'ğŠ'

LinearBSyllableB066Ta2

\u{1004b}: 'ğ‹'

LinearBSyllableB087Twe

\u{1004c}: 'ğŒ'

LinearBSyllableB091Two

\u{1004d}: 'ğ'

LinearBSymbolB018

\u{10050}: 'ğ'

LinearBSymbolB019

\u{10051}: 'ğ‘'

LinearBSymbolB022

\u{10052}: 'ğ’'

LinearBSymbolB034

\u{10053}: 'ğ“'

LinearBSymbolB047

\u{10054}: 'ğ”'

LinearBSymbolB049

\u{10055}: 'ğ•'

LinearBSymbolB056

\u{10056}: 'ğ–'

LinearBSymbolB063

\u{10057}: 'ğ—'

LinearBSymbolB064

\u{10058}: 'ğ˜'

LinearBSymbolB079

\u{10059}: 'ğ™'

LinearBSymbolB082

\u{1005a}: 'ğš'

LinearBSymbolB083

\u{1005b}: 'ğ›'

LinearBSymbolB086

\u{1005c}: 'ğœ'

LinearBSymbolB089

\u{1005d}: 'ğ'

Methods

impl LinearBSyllabary

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case