Enum unicode_types::mende_kikakui::MendeKikakui

pub enum MendeKikakui {
    SyllableM001Ki,
    SyllableM002Ka,
    SyllableM003Ku,
    SyllableM065Kee,
    SyllableM095Ke,
    SyllableM076Koo,
    SyllableM048Ko,
    SyllableM179Kua,
    SyllableM004Wi,
    SyllableM005Wa,
    SyllableM006Wu,
    SyllableM126Wee,
    SyllableM118We,
    SyllableM114Woo,
    SyllableM045Wo,
    SyllableM194Wui,
    SyllableM143Wei,
    SyllableM061Wvi,
    SyllableM049Wva,
    SyllableM139Wve,
    SyllableM007Min,
    SyllableM008Man,
    SyllableM009Mun,
    SyllableM059Men,
    SyllableM094Mon,
    SyllableM154Muan,
    SyllableM189Muen,
    SyllableM010Bi,
    SyllableM011Ba,
    SyllableM012Bu,
    SyllableM150Bee,
    SyllableM097Be,
    SyllableM103Boo,
    SyllableM138Bo,
    SyllableM013I,
    SyllableM014A,
    SyllableM015U,
    SyllableM163Ee,
    SyllableM100E,
    SyllableM165Oo,
    SyllableM147O,
    SyllableM137Ei,
    SyllableM131In,
    SyllableM135In,
    SyllableM195An,
    SyllableM178En,
    SyllableM019Si,
    SyllableM020Sa,
    SyllableM021Su,
    SyllableM162See,
    SyllableM116Se,
    SyllableM136Soo,
    SyllableM079So,
    SyllableM196Sia,
    SyllableM025Li,
    SyllableM026La,
    SyllableM027Lu,
    SyllableM084Lee,
    SyllableM073Le,
    SyllableM054Loo,
    SyllableM153Lo,
    SyllableM110LongLe,
    SyllableM016Di,
    SyllableM017Da,
    SyllableM018Du,
    SyllableM089Dee,
    SyllableM180Doo,
    SyllableM181Do,
    SyllableM022Ti,
    SyllableM023Ta,
    SyllableM024Tu,
    SyllableM091Tee,
    SyllableM055Te,
    SyllableM104Too,
    SyllableM069To,
    SyllableM028Ji,
    SyllableM029Ja,
    SyllableM030Ju,
    SyllableM157Jee,
    SyllableM113Je,
    SyllableM160Joo,
    SyllableM063Jo,
    SyllableM175LongJo,
    SyllableM031Yi,
    SyllableM032Ya,
    SyllableM033Yu,
    SyllableM109Yee,
    SyllableM080Ye,
    SyllableM141Yoo,
    SyllableM121Yo,
    SyllableM034Fi,
    SyllableM035Fa,
    SyllableM036Fu,
    SyllableM078Fee,
    SyllableM075Fe,
    SyllableM133Foo,
    SyllableM088Fo,
    SyllableM197Fua,
    SyllableM101Fan,
    SyllableM037Nin,
    SyllableM038Nan,
    SyllableM039Nun,
    SyllableM117Nen,
    SyllableM169Non,
    SyllableM176Hi,
    SyllableM041Ha,
    SyllableM186Hu,
    SyllableM040Hee,
    SyllableM096He,
    SyllableM042Hoo,
    SyllableM140Ho,
    SyllableM083Heei,
    SyllableM128Hoou,
    SyllableM053Hin,
    SyllableM130Han,
    SyllableM087Hun,
    SyllableM052Hen,
    SyllableM193Hon,
    SyllableM046Huan,
    SyllableM090Nggi,
    SyllableM043Ngga,
    SyllableM082Nggu,
    SyllableM115Nggee,
    SyllableM146Ngge,
    SyllableM156Nggoo,
    SyllableM120Nggo,
    SyllableM159Nggaa,
    SyllableM127Nggua,
    SyllableM086LongNgge,
    SyllableM106LongNggoo,
    SyllableM183LongNggo,
    SyllableM155Gi,
    SyllableM111Ga,
    SyllableM168Gu,
    SyllableM190Gee,
    SyllableM166Guei,
    SyllableM167Guan,
    SyllableM184Ngen,
    SyllableM057Ngon,
    SyllableM177Nguan,
    SyllableM068Pi,
    SyllableM099Pa,
    SyllableM050Pu,
    SyllableM081Pee,
    SyllableM051Pe,
    SyllableM102Poo,
    SyllableM066Po,
    SyllableM145Mbi,
    SyllableM062Mba,
    SyllableM122Mbu,
    SyllableM047Mbee,
    SyllableM188Mbee,
    SyllableM072Mbe,
    SyllableM172Mboo,
    SyllableM174Mbo,
    SyllableM187Mbuu,
    SyllableM161LongMbe,
    SyllableM105LongMboo,
    SyllableM142LongMbo,
    SyllableM132Kpi,
    SyllableM092Kpa,
    SyllableM074Kpu,
    SyllableM044Kpee,
    SyllableM108Kpe,
    SyllableM112Kpoo,
    SyllableM158Kpo,
    SyllableM124Gbi,
    SyllableM056Gba,
    SyllableM148Gbu,
    SyllableM093Gbee,
    SyllableM107Gbe,
    SyllableM071Gboo,
    SyllableM070Gbo,
    SyllableM171Ra,
    SyllableM123Ndi,
    SyllableM129Nda,
    SyllableM125Ndu,
    SyllableM191Ndee,
    SyllableM119Nde,
    SyllableM067Ndoo,
    SyllableM064Ndo,
    SyllableM152Nja,
    SyllableM192Nju,
    SyllableM149Njee,
    SyllableM134Njoo,
    SyllableM182Vi,
    SyllableM185Va,
    SyllableM151Vu,
    SyllableM173Vee,
    SyllableM085Ve,
    SyllableM144Voo,
    SyllableM077Vo,
    SyllableM164Nyin,
    SyllableM058Nyan,
    SyllableM170Nyun,
    SyllableM098Nyen,
    SyllableM060Nyon,
    DigitOne,
    DigitTwo,
    DigitThree,
    DigitFour,
    DigitFive,
    DigitSix,
    DigitSeven,
    DigitEight,
    DigitNine,
    CombiningNumberTeens,
    CombiningNumberTens,
    CombiningNumberHundreds,
    CombiningNumberThousands,
    CombiningNumberTenThousands,
    CombiningNumberHundredThousands,
    CombiningNumberMillions,
}

An enum to represent all characters in the MendeKikakui block.

Variants
SyllableM001Ki

\u{1e800}: 'ğ €'

SyllableM002Ka

\u{1e801}: 'ğ '

SyllableM003Ku

\u{1e802}: 'ğ ‚'

SyllableM065Kee

\u{1e803}: 'ğ ƒ'

SyllableM095Ke

\u{1e804}: 'ğ „'

SyllableM076Koo

\u{1e805}: 'ğ …'

SyllableM048Ko

\u{1e806}: 'ğ †'

SyllableM179Kua

\u{1e807}: 'ğ ‡'

SyllableM004Wi

\u{1e808}: 'ğ ˆ'

SyllableM005Wa

\u{1e809}: 'ğ ‰'

SyllableM006Wu

\u{1e80a}: 'ğ Š'

SyllableM126Wee

\u{1e80b}: 'ğ ‹'

SyllableM118We

\u{1e80c}: 'ğ Œ'

SyllableM114Woo

\u{1e80d}: 'ğ '

SyllableM045Wo

\u{1e80e}: 'ğ '

SyllableM194Wui

\u{1e80f}: 'ğ '

SyllableM143Wei

\u{1e810}: 'ğ '

SyllableM061Wvi

\u{1e811}: 'ğ ‘'

SyllableM049Wva

\u{1e812}: 'ğ ’'

SyllableM139Wve

\u{1e813}: 'ğ “'

SyllableM007Min

\u{1e814}: 'ğ ”'

SyllableM008Man

\u{1e815}: 'ğ •'

SyllableM009Mun

\u{1e816}: 'ğ –'

SyllableM059Men

\u{1e817}: 'ğ —'

SyllableM094Mon

\u{1e818}: 'ğ ˜'

SyllableM154Muan

\u{1e819}: 'ğ ™'

SyllableM189Muen

\u{1e81a}: 'ğ š'

SyllableM010Bi

\u{1e81b}: 'ğ ›'

SyllableM011Ba

\u{1e81c}: 'ğ œ'

SyllableM012Bu

\u{1e81d}: 'ğ '

SyllableM150Bee

\u{1e81e}: 'ğ '

SyllableM097Be

\u{1e81f}: 'ğ Ÿ'

SyllableM103Boo

\u{1e820}: 'ğ '

SyllableM138Bo

\u{1e821}: 'ğ ¡'

SyllableM013I

\u{1e822}: 'ğ ¢'

SyllableM014A

\u{1e823}: 'ğ £'

SyllableM015U

\u{1e824}: 'ğ ¤'

SyllableM163Ee

\u{1e825}: 'ğ ¥'

SyllableM100E

\u{1e826}: 'ğ ¦'

SyllableM165Oo

\u{1e827}: 'ğ '

SyllableM147O

\u{1e828}: 'ğ ¨'

SyllableM137Ei

\u{1e829}: 'ğ ©'

SyllableM131In

\u{1e82a}: 'ğ ª'

SyllableM135In

\u{1e82b}: 'ğ «'

SyllableM195An

\u{1e82c}: 'ğ ¬'

SyllableM178En

\u{1e82d}: 'ğ ­'

SyllableM019Si

\u{1e82e}: 'ğ ®'

SyllableM020Sa

\u{1e82f}: 'ğ ¯'

SyllableM021Su

\u{1e830}: 'ğ °'

SyllableM162See

\u{1e831}: 'ğ ±'

SyllableM116Se

\u{1e832}: 'ğ ²'

SyllableM136Soo

\u{1e833}: 'ğ ³'

SyllableM079So

\u{1e834}: 'ğ ´'

SyllableM196Sia

\u{1e835}: 'ğ µ'

SyllableM025Li

\u{1e836}: 'ğ ¶'

SyllableM026La

\u{1e837}: 'ğ ·'

SyllableM027Lu

\u{1e838}: 'ğ ¸'

SyllableM084Lee

\u{1e839}: 'ğ ¹'

SyllableM073Le

\u{1e83a}: 'ğ º'

SyllableM054Loo

\u{1e83b}: 'ğ »'

SyllableM153Lo

\u{1e83c}: 'ğ ¼'

SyllableM110LongLe

\u{1e83d}: 'ğ ½'

SyllableM016Di

\u{1e83e}: 'ğ ¾'

SyllableM017Da

\u{1e83f}: 'ğ ¿'

SyllableM018Du

\u{1e840}: 'ğ¡€'

SyllableM089Dee

\u{1e841}: 'ğ¡'

SyllableM180Doo

\u{1e842}: 'ğ¡‚'

SyllableM181Do

\u{1e843}: 'ğ¡ƒ'

SyllableM022Ti

\u{1e844}: 'ğ¡„'

SyllableM023Ta

\u{1e845}: 'ğ¡…'

SyllableM024Tu

\u{1e846}: 'ğ¡†'

SyllableM091Tee

\u{1e847}: 'ğ¡‡'

SyllableM055Te

\u{1e848}: 'ğ¡ˆ'

SyllableM104Too

\u{1e849}: 'ğ¡‰'

SyllableM069To

\u{1e84a}: 'ğ¡Š'

SyllableM028Ji

\u{1e84b}: 'ğ¡‹'

SyllableM029Ja

\u{1e84c}: 'ğ¡Œ'

SyllableM030Ju

\u{1e84d}: 'ğ¡'

SyllableM157Jee

\u{1e84e}: 'ğ¡'

SyllableM113Je

\u{1e84f}: 'ğ¡'

SyllableM160Joo

\u{1e850}: 'ğ¡'

SyllableM063Jo

\u{1e851}: 'ğ¡‘'

SyllableM175LongJo

\u{1e852}: 'ğ¡’'

SyllableM031Yi

\u{1e853}: 'ğ¡“'

SyllableM032Ya

\u{1e854}: 'ğ¡”'

SyllableM033Yu

\u{1e855}: 'ğ¡•'

SyllableM109Yee

\u{1e856}: 'ğ¡–'

SyllableM080Ye

\u{1e857}: 'ğ¡—'

SyllableM141Yoo

\u{1e858}: 'ğ¡˜'

SyllableM121Yo

\u{1e859}: 'ğ¡™'

SyllableM034Fi

\u{1e85a}: 'ğ¡š'

SyllableM035Fa

\u{1e85b}: 'ğ¡›'

SyllableM036Fu

\u{1e85c}: 'ğ¡œ'

SyllableM078Fee

\u{1e85d}: 'ğ¡'

SyllableM075Fe

\u{1e85e}: 'ğ¡'

SyllableM133Foo

\u{1e85f}: 'ğ¡Ÿ'

SyllableM088Fo

\u{1e860}: 'ğ¡ '

SyllableM197Fua

\u{1e861}: 'ğ¡¡'

SyllableM101Fan

\u{1e862}: 'ğ¡¢'

SyllableM037Nin

\u{1e863}: 'ğ¡£'

SyllableM038Nan

\u{1e864}: 'ğ¡¤'

SyllableM039Nun

\u{1e865}: 'ğ¡¥'

SyllableM117Nen

\u{1e866}: 'ğ¡¦'

SyllableM169Non

\u{1e867}: 'ğ¡'

SyllableM176Hi

\u{1e868}: 'ğ¡¨'

SyllableM041Ha

\u{1e869}: 'ğ¡©'

SyllableM186Hu

\u{1e86a}: 'ğ¡ª'

SyllableM040Hee

\u{1e86b}: 'ğ¡«'

SyllableM096He

\u{1e86c}: 'ğ¡¬'

SyllableM042Hoo

\u{1e86d}: 'ğ¡­'

SyllableM140Ho

\u{1e86e}: 'ğ¡®'

SyllableM083Heei

\u{1e86f}: 'ğ¡¯'

SyllableM128Hoou

\u{1e870}: 'ğ¡°'

SyllableM053Hin

\u{1e871}: 'ğ¡±'

SyllableM130Han

\u{1e872}: 'ğ¡²'

SyllableM087Hun

\u{1e873}: 'ğ¡³'

SyllableM052Hen

\u{1e874}: 'ğ¡´'

SyllableM193Hon

\u{1e875}: 'ğ¡µ'

SyllableM046Huan

\u{1e876}: 'ğ¡¶'

SyllableM090Nggi

\u{1e877}: 'ğ¡·'

SyllableM043Ngga

\u{1e878}: 'ğ¡¸'

SyllableM082Nggu

\u{1e879}: 'ğ¡¹'

SyllableM115Nggee

\u{1e87a}: 'ğ¡º'

SyllableM146Ngge

\u{1e87b}: 'ğ¡»'

SyllableM156Nggoo

\u{1e87c}: 'ğ¡¼'

SyllableM120Nggo

\u{1e87d}: 'ğ¡½'

SyllableM159Nggaa

\u{1e87e}: 'ğ¡¾'

SyllableM127Nggua

\u{1e87f}: 'ğ¡¿'

SyllableM086LongNgge

\u{1e880}: 'ğ¢€'

SyllableM106LongNggoo

\u{1e881}: 'ğ¢'

SyllableM183LongNggo

\u{1e882}: 'ğ¢‚'

SyllableM155Gi

\u{1e883}: 'ğ¢ƒ'

SyllableM111Ga

\u{1e884}: 'ğ¢„'

SyllableM168Gu

\u{1e885}: 'ğ¢…'

SyllableM190Gee

\u{1e886}: 'ğ¢†'

SyllableM166Guei

\u{1e887}: 'ğ¢‡'

SyllableM167Guan

\u{1e888}: 'ğ¢ˆ'

SyllableM184Ngen

\u{1e889}: 'ğ¢‰'

SyllableM057Ngon

\u{1e88a}: 'ğ¢Š'

SyllableM177Nguan

\u{1e88b}: 'ğ¢‹'

SyllableM068Pi

\u{1e88c}: 'ğ¢Œ'

SyllableM099Pa

\u{1e88d}: 'ğ¢'

SyllableM050Pu

\u{1e88e}: 'ğ¢'

SyllableM081Pee

\u{1e88f}: 'ğ¢'

SyllableM051Pe

\u{1e890}: 'ğ¢'

SyllableM102Poo

\u{1e891}: 'ğ¢‘'

SyllableM066Po

\u{1e892}: 'ğ¢’'

SyllableM145Mbi

\u{1e893}: 'ğ¢“'

SyllableM062Mba

\u{1e894}: 'ğ¢”'

SyllableM122Mbu

\u{1e895}: 'ğ¢•'

SyllableM047Mbee

\u{1e896}: 'ğ¢–'

SyllableM188Mbee

\u{1e897}: 'ğ¢—'

SyllableM072Mbe

\u{1e898}: 'ğ¢˜'

SyllableM172Mboo

\u{1e899}: 'ğ¢™'

SyllableM174Mbo

\u{1e89a}: 'ğ¢š'

SyllableM187Mbuu

\u{1e89b}: 'ğ¢›'

SyllableM161LongMbe

\u{1e89c}: 'ğ¢œ'

SyllableM105LongMboo

\u{1e89d}: 'ğ¢'

SyllableM142LongMbo

\u{1e89e}: 'ğ¢'

SyllableM132Kpi

\u{1e89f}: 'ğ¢Ÿ'

SyllableM092Kpa

\u{1e8a0}: 'ğ¢ '

SyllableM074Kpu

\u{1e8a1}: 'ğ¢¡'

SyllableM044Kpee

\u{1e8a2}: 'ğ¢¢'

SyllableM108Kpe

\u{1e8a3}: 'ğ¢£'

SyllableM112Kpoo

\u{1e8a4}: 'ğ¢¤'

SyllableM158Kpo

\u{1e8a5}: 'ğ¢¥'

SyllableM124Gbi

\u{1e8a6}: 'ğ¢¦'

SyllableM056Gba

\u{1e8a7}: 'ğ¢'

SyllableM148Gbu

\u{1e8a8}: 'ğ¢¨'

SyllableM093Gbee

\u{1e8a9}: 'ğ¢©'

SyllableM107Gbe

\u{1e8aa}: 'ğ¢ª'

SyllableM071Gboo

\u{1e8ab}: 'ğ¢«'

SyllableM070Gbo

\u{1e8ac}: 'ğ¢¬'

SyllableM171Ra

\u{1e8ad}: 'ğ¢­'

SyllableM123Ndi

\u{1e8ae}: 'ğ¢®'

SyllableM129Nda

\u{1e8af}: 'ğ¢¯'

SyllableM125Ndu

\u{1e8b0}: 'ğ¢°'

SyllableM191Ndee

\u{1e8b1}: 'ğ¢±'

SyllableM119Nde

\u{1e8b2}: 'ğ¢²'

SyllableM067Ndoo

\u{1e8b3}: 'ğ¢³'

SyllableM064Ndo

\u{1e8b4}: 'ğ¢´'

SyllableM152Nja

\u{1e8b5}: 'ğ¢µ'

SyllableM192Nju

\u{1e8b6}: 'ğ¢¶'

SyllableM149Njee

\u{1e8b7}: 'ğ¢·'

SyllableM134Njoo

\u{1e8b8}: 'ğ¢¸'

SyllableM182Vi

\u{1e8b9}: 'ğ¢¹'

SyllableM185Va

\u{1e8ba}: 'ğ¢º'

SyllableM151Vu

\u{1e8bb}: 'ğ¢»'

SyllableM173Vee

\u{1e8bc}: 'ğ¢¼'

SyllableM085Ve

\u{1e8bd}: 'ğ¢½'

SyllableM144Voo

\u{1e8be}: 'ğ¢¾'

SyllableM077Vo

\u{1e8bf}: 'ğ¢¿'

SyllableM164Nyin

\u{1e8c0}: 'ğ£€'

SyllableM058Nyan

\u{1e8c1}: 'ğ£'

SyllableM170Nyun

\u{1e8c2}: 'ğ£‚'

SyllableM098Nyen

\u{1e8c3}: 'ğ£ƒ'

SyllableM060Nyon

\u{1e8c4}: 'ğ£„'

DigitOne

\u{1e8c7}: 'ğ£‡'

DigitTwo

\u{1e8c8}: 'ğ£ˆ'

DigitThree

\u{1e8c9}: 'ğ£‰'

DigitFour

\u{1e8ca}: 'ğ£Š'

DigitFive

\u{1e8cb}: 'ğ£‹'

DigitSix

\u{1e8cc}: 'ğ£Œ'

DigitSeven

\u{1e8cd}: 'ğ£'

DigitEight

\u{1e8ce}: 'ğ£'

DigitNine

\u{1e8cf}: 'ğ£'

CombiningNumberTeens

\u{1e8d0}: 'ğ£'

CombiningNumberTens

\u{1e8d1}: 'ğ£‘'

CombiningNumberHundreds

\u{1e8d2}: 'ğ£’'

CombiningNumberThousands

\u{1e8d3}: 'ğ£“'

CombiningNumberTenThousands

\u{1e8d4}: 'ğ£”'

CombiningNumberHundredThousands

\u{1e8d5}: 'ğ£•'

CombiningNumberMillions

\u{1e8d6}: 'ğ£–'

Methods

impl MendeKikakui

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case