Enum unicode_types::meroitic_cursive::MeroiticCursive

pub enum MeroiticCursive {
    LetterA,
    LetterE,
    LetterI,
    LetterO,
    LetterYa,
    LetterWa,
    LetterBa,
    LetterPa,
    LetterMa,
    LetterNa,
    LetterNe,
    LetterRa,
    LetterLa,
    LetterKha,
    LetterHha,
    LetterSa,
    LetterArchaicSa,
    LetterSe,
    LetterKa,
    LetterQa,
    LetterTa,
    LetterTe,
    LetterTo,
    LetterDa,
    FractionElevenTwelfths,
    FractionOneHalf,
    LogogramRmt,
    LogogramImn,
    NumberOne,
    NumberTwo,
    NumberThree,
    NumberFour,
    NumberFive,
    NumberSix,
    NumberSeven,
    NumberEight,
    NumberNine,
    NumberTen,
    NumberTwenty,
    NumberThirty,
    NumberForty,
    NumberFifty,
    NumberSixty,
    NumberSeventy,
    NumberOneHundred,
    NumberTwoHundred,
    NumberThreeHundred,
    NumberFourHundred,
    NumberFiveHundred,
    NumberSixHundred,
    NumberSevenHundred,
    NumberEightHundred,
    NumberNineHundred,
    NumberOneThousand,
    NumberTwoThousand,
    NumberThreeThousand,
    NumberFourThousand,
    NumberFiveThousand,
    NumberSixThousand,
    NumberSevenThousand,
    NumberEightThousand,
    NumberNineThousand,
    NumberTenThousand,
    NumberTwentyThousand,
    NumberThirtyThousand,
    NumberFortyThousand,
    NumberFiftyThousand,
    NumberSixtyThousand,
    NumberSeventyThousand,
    NumberEightyThousand,
    NumberNinetyThousand,
    NumberOneHundredThousand,
    NumberTwoHundredThousand,
    NumberThreeHundredThousand,
    NumberFourHundredThousand,
    NumberFiveHundredThousand,
    NumberSixHundredThousand,
    NumberSevenHundredThousand,
    NumberEightHundredThousand,
    NumberNineHundredThousand,
    FractionOneTwelfth,
    FractionTwoTwelfths,
    FractionThreeTwelfths,
    FractionFourTwelfths,
    FractionFiveTwelfths,
    FractionSixTwelfths,
    FractionSevenTwelfths,
    FractionEightTwelfths,
    FractionNineTwelfths,
}

An enum to represent all characters in the MeroiticCursive block.

Variants
LetterA

\u{109a0}: 'ğ¦ '

LetterE

\u{109a1}: 'ğ¦¡'

LetterI

\u{109a2}: 'ğ¦¢'

LetterO

\u{109a3}: 'ğ¦£'

LetterYa

\u{109a4}: 'ğ¦¤'

LetterWa

\u{109a5}: 'ğ¦¥'

LetterBa

\u{109a6}: 'ğ¦¦'

LetterPa

\u{109a7}: 'ğ¦'

LetterMa

\u{109a8}: 'ğ¦¨'

LetterNa

\u{109a9}: 'ğ¦©'

LetterNe

\u{109aa}: 'ğ¦ª'

LetterRa

\u{109ab}: 'ğ¦«'

LetterLa

\u{109ac}: 'ğ¦¬'

LetterKha

\u{109ad}: 'ğ¦­'

LetterHha

\u{109ae}: 'ğ¦®'

LetterSa

\u{109af}: 'ğ¦¯'

LetterArchaicSa

\u{109b0}: 'ğ¦°'

LetterSe

\u{109b1}: 'ğ¦±'

LetterKa

\u{109b2}: 'ğ¦²'

LetterQa

\u{109b3}: 'ğ¦³'

LetterTa

\u{109b4}: 'ğ¦´'

LetterTe

\u{109b5}: 'ğ¦µ'

LetterTo

\u{109b6}: 'ğ¦¶'

LetterDa

\u{109b7}: 'ğ¦·'

FractionElevenTwelfths

\u{109bc}: 'ğ¦¼'

FractionOneHalf

\u{109bd}: 'ğ¦½'

LogogramRmt

\u{109be}: 'ğ¦¾'

LogogramImn

\u{109bf}: 'ğ¦¿'

NumberOne

\u{109c0}: 'ğ€'

NumberTwo

\u{109c1}: 'ğ'

NumberThree

\u{109c2}: 'ğ‚'

NumberFour

\u{109c3}: 'ğƒ'

NumberFive

\u{109c4}: 'ğ„'

NumberSix

\u{109c5}: 'ğ…'

NumberSeven

\u{109c6}: 'ğ†'

NumberEight

\u{109c7}: 'ğ‡'

NumberNine

\u{109c8}: 'ğˆ'

NumberTen

\u{109c9}: 'ğ‰'

NumberTwenty

\u{109ca}: 'ğŠ'

NumberThirty

\u{109cb}: 'ğ‹'

NumberForty

\u{109cc}: 'ğŒ'

NumberFifty

\u{109cd}: 'ğ'

NumberSixty

\u{109ce}: 'ğ'

NumberSeventy

\u{109cf}: 'ğ'

NumberOneHundred

\u{109d2}: 'ğ’'

NumberTwoHundred

\u{109d3}: 'ğ“'

NumberThreeHundred

\u{109d4}: 'ğ”'

NumberFourHundred

\u{109d5}: 'ğ•'

NumberFiveHundred

\u{109d6}: 'ğ–'

NumberSixHundred

\u{109d7}: 'ğ—'

NumberSevenHundred

\u{109d8}: 'ğ˜'

NumberEightHundred

\u{109d9}: 'ğ™'

NumberNineHundred

\u{109da}: 'ğš'

NumberOneThousand

\u{109db}: 'ğ›'

NumberTwoThousand

\u{109dc}: 'ğœ'

NumberThreeThousand

\u{109dd}: 'ğ'

NumberFourThousand

\u{109de}: 'ğ'

NumberFiveThousand

\u{109df}: 'ğŸ'

NumberSixThousand

\u{109e0}: 'ğ '

NumberSevenThousand

\u{109e1}: 'ğ¡'

NumberEightThousand

\u{109e2}: 'ğ¢'

NumberNineThousand

\u{109e3}: 'ğ£'

NumberTenThousand

\u{109e4}: 'ğ¤'

NumberTwentyThousand

\u{109e5}: 'ğ¥'

NumberThirtyThousand

\u{109e6}: 'ğ¦'

NumberFortyThousand

\u{109e7}: 'ğ'

NumberFiftyThousand

\u{109e8}: 'ğ¨'

NumberSixtyThousand

\u{109e9}: 'ğ©'

NumberSeventyThousand

\u{109ea}: 'ğª'

NumberEightyThousand

\u{109eb}: 'ğ«'

NumberNinetyThousand

\u{109ec}: 'ğ¬'

NumberOneHundredThousand

\u{109ed}: 'ğ­'

NumberTwoHundredThousand

\u{109ee}: 'ğ®'

NumberThreeHundredThousand

\u{109ef}: 'ğ¯'

NumberFourHundredThousand

\u{109f0}: 'ğ°'

NumberFiveHundredThousand

\u{109f1}: 'ğ±'

NumberSixHundredThousand

\u{109f2}: 'ğ²'

NumberSevenHundredThousand

\u{109f3}: 'ğ³'

NumberEightHundredThousand

\u{109f4}: 'ğ´'

NumberNineHundredThousand

\u{109f5}: 'ğµ'

FractionOneTwelfth

\u{109f6}: 'ğ¶'

FractionTwoTwelfths

\u{109f7}: 'ğ·'

FractionThreeTwelfths

\u{109f8}: 'ğ¸'

FractionFourTwelfths

\u{109f9}: 'ğ¹'

FractionFiveTwelfths

\u{109fa}: 'ğº'

FractionSixTwelfths

\u{109fb}: 'ğ»'

FractionSevenTwelfths

\u{109fc}: 'ğ¼'

FractionEightTwelfths

\u{109fd}: 'ğ½'

FractionNineTwelfths

\u{109fe}: 'ğ¾'

Methods

impl MeroiticCursive

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case