Enum unicode_types::miscellaneous_mathematical_symbols_b::MiscellaneousMathematicalSymbolsB

pub enum MiscellaneousMathematicalSymbolsB {
    TripleVerticalBarDelimiter,
    ZNotationSpot,
    ZNotationTypeColon,
    LeftWhiteCurlyBracket,
    RightWhiteCurlyBracket,
    LeftWhiteParenthesis,
    RightWhiteParenthesis,
    ZNotationLeftImageBracket,
    ZNotationRightImageBracket,
    ZNotationLeftBindingBracket,
    ZNotationRightBindingBracket,
    LeftSquareBracketWithUnderbar,
    RightSquareBracketWithUnderbar,
    LeftSquareBracketWithTickInTopCorner,
    RightSquareBracketWithTickInBottomCorner,
    LeftSquareBracketWithTickInBottomCorner,
    RightSquareBracketWithTickInTopCorner,
    LeftAngleBracketWithDot,
    RightAngleBracketWithDot,
    LeftArcLessDashThanBracket,
    RightArcGreaterDashThanBracket,
    DoubleLeftArcGreaterDashThanBracket,
    DoubleRightArcLessDashThanBracket,
    LeftBlackTortoiseShellBracket,
    RightBlackTortoiseShellBracket,
    DottedFence,
    VerticalZigzagLine,
    MeasuredAngleOpeningLeft,
    RightAngleVariantWithSquare,
    MeasuredRightAngleWithDot,
    AngleWithSInside,
    AcuteAngle,
    SphericalAngleOpeningLeft,
    SphericalAngleOpeningUp,
    TurnedAngle,
    ReversedAngle,
    AngleWithUnderbar,
    ReversedAngleWithUnderbar,
    ObliqueAngleOpeningUp,
    ObliqueAngleOpeningDown,
    MeasuredAngleWithOpenArmEndingInArrowPointingUpAndRight,
    MeasuredAngleWithOpenArmEndingInArrowPointingUpAndLeft,
    MeasuredAngleWithOpenArmEndingInArrowPointingDownAndRight,
    MeasuredAngleWithOpenArmEndingInArrowPointingDownAndLeft,
    MeasuredAngleWithOpenArmEndingInArrowPointingRightAndUp,
    MeasuredAngleWithOpenArmEndingInArrowPointingLeftAndUp,
    MeasuredAngleWithOpenArmEndingInArrowPointingRightAndDown,
    MeasuredAngleWithOpenArmEndingInArrowPointingLeftAndDown,
    ReversedEmptySet,
    EmptySetWithOverbar,
    EmptySetWithSmallCircleAbove,
    EmptySetWithRightArrowAbove,
    EmptySetWithLeftArrowAbove,
    CircleWithHorizontalBar,
    CircledVerticalBar,
    CircledParallel,
    CircledReverseSolidus,
    CircledPerpendicular,
    CircleDividedByHorizontalBarAndTopHalfDividedByVerticalBar,
    CircleWithSuperimposedX,
    CircledAnticlockwiseDashRotatedDivisionSign,
    UpArrowThroughCircle,
    CircledWhiteBullet,
    CircledBullet,
    CircledLessDashThan,
    CircledGreaterDashThan,
    CircleWithSmallCircleToTheRight,
    CircleWithTwoHorizontalStrokesToTheRight,
    SquaredRisingDiagonalSlash,
    SquaredFallingDiagonalSlash,
    SquaredAsterisk,
    SquaredSmallCircle,
    SquaredSquare,
    TwoJoinedSquares,
    TriangleWithDotAbove,
    TriangleWithUnderbar,
    SInTriangle,
    TriangleWithSerifsAtBottom,
    RightTriangleAboveLeftTriangle,
    LeftTriangleBesideVerticalBar,
    VerticalBarBesideRightTriangle,
    BowtieWithLeftHalfBlack,
    BowtieWithRightHalfBlack,
    BlackBowtie,
    TimesWithLeftHalfBlack,
    TimesWithRightHalfBlack,
    WhiteHourglass,
    BlackHourglass,
    LeftWigglyFence,
    RightWigglyFence,
    LeftDoubleWigglyFence,
    RightDoubleWigglyFence,
    IncompleteInfinity,
    TieOverInfinity,
    InfinityNegatedWithVerticalBar,
    DoubleDashEndedMultimap,
    SquareWithContouredOutline,
    IncreasesAs,
    ShuffleProduct,
    EqualsSignAndSlantedParallel,
    EqualsSignAndSlantedParallelWithTildeAbove,
    IdenticalToAndSlantedParallel,
    GleichStark,
    Thermodynamic,
    DownDashPointingTriangleWithLeftHalfBlack,
    DownDashPointingTriangleWithRightHalfBlack,
    BlackDiamondWithDownArrow,
    BlackLozenge,
    WhiteCircleWithDownArrow,
    BlackCircleWithDownArrow,
    ErrorDashBarredWhiteSquare,
    ErrorDashBarredBlackSquare,
    ErrorDashBarredWhiteDiamond,
    ErrorDashBarredBlackDiamond,
    ErrorDashBarredWhiteCircle,
    ErrorDashBarredBlackCircle,
    RuleDashDelayed,
    ReverseSolidusOperator,
    SolidusWithOverbar,
    ReverseSolidusWithHorizontalStroke,
    BigSolidus,
    BigReverseSolidus,
    DoublePlus,
    TriplePlus,
    LeftDashPointingCurvedAngleBracket,
    RightDashPointingCurvedAngleBracket,
    Tiny,
}

An enum to represent all characters in the MiscellaneousMathematicalSymbolsB block.

Variants
TripleVerticalBarDelimiter

\u{2980}: '⦀'

ZNotationSpot

\u{2981}: '⦁'

ZNotationTypeColon

\u{2982}: '⦂'

LeftWhiteCurlyBracket

\u{2983}: '⦃'

RightWhiteCurlyBracket

\u{2984}: '⦄'

LeftWhiteParenthesis

\u{2985}: '⦅'

RightWhiteParenthesis

\u{2986}: '⦆'

ZNotationLeftImageBracket

\u{2987}: '⦇'

ZNotationRightImageBracket

\u{2988}: '⦈'

ZNotationLeftBindingBracket

\u{2989}: '⦉'

ZNotationRightBindingBracket

\u{298a}: '⦊'

LeftSquareBracketWithUnderbar

\u{298b}: '⦋'

RightSquareBracketWithUnderbar

\u{298c}: '⦌'

LeftSquareBracketWithTickInTopCorner

\u{298d}: '⦍'

RightSquareBracketWithTickInBottomCorner

\u{298e}: '⦎'

LeftSquareBracketWithTickInBottomCorner

\u{298f}: '⦏'

RightSquareBracketWithTickInTopCorner

\u{2990}: '⦐'

LeftAngleBracketWithDot

\u{2991}: '⦑'

RightAngleBracketWithDot

\u{2992}: '⦒'

LeftArcLessDashThanBracket

\u{2993}: '⦓'

RightArcGreaterDashThanBracket

\u{2994}: '⦔'

DoubleLeftArcGreaterDashThanBracket

\u{2995}: '⦕'

DoubleRightArcLessDashThanBracket

\u{2996}: '⦖'

LeftBlackTortoiseShellBracket

\u{2997}: '⦗'

RightBlackTortoiseShellBracket

\u{2998}: '⦘'

DottedFence

\u{2999}: '⦙'

VerticalZigzagLine

\u{299a}: '⦚'

MeasuredAngleOpeningLeft

\u{299b}: '⦛'

RightAngleVariantWithSquare

\u{299c}: '⦜'

MeasuredRightAngleWithDot

\u{299d}: '⦝'

AngleWithSInside

\u{299e}: '⦞'

AcuteAngle

\u{299f}: '⦟'

SphericalAngleOpeningLeft

\u{29a0}: '⦠'

SphericalAngleOpeningUp

\u{29a1}: '⦡'

TurnedAngle

\u{29a2}: '⦢'

ReversedAngle

\u{29a3}: '⦣'

AngleWithUnderbar

\u{29a4}: '⦤'

ReversedAngleWithUnderbar

\u{29a5}: '⦥'

ObliqueAngleOpeningUp

\u{29a6}: '⦦'

ObliqueAngleOpeningDown

\u{29a7}: '⦧'

MeasuredAngleWithOpenArmEndingInArrowPointingUpAndRight

\u{29a8}: '⦨'

MeasuredAngleWithOpenArmEndingInArrowPointingUpAndLeft

\u{29a9}: '⦩'

MeasuredAngleWithOpenArmEndingInArrowPointingDownAndRight

\u{29aa}: '⦪'

MeasuredAngleWithOpenArmEndingInArrowPointingDownAndLeft

\u{29ab}: '⦫'

MeasuredAngleWithOpenArmEndingInArrowPointingRightAndUp

\u{29ac}: '⦬'

MeasuredAngleWithOpenArmEndingInArrowPointingLeftAndUp

\u{29ad}: '⦭'

MeasuredAngleWithOpenArmEndingInArrowPointingRightAndDown

\u{29ae}: '⦮'

MeasuredAngleWithOpenArmEndingInArrowPointingLeftAndDown

\u{29af}: '⦯'

ReversedEmptySet

\u{29b0}: '⦰'

EmptySetWithOverbar

\u{29b1}: '⦱'

EmptySetWithSmallCircleAbove

\u{29b2}: '⦲'

EmptySetWithRightArrowAbove

\u{29b3}: '⦳'

EmptySetWithLeftArrowAbove

\u{29b4}: '⦴'

CircleWithHorizontalBar

\u{29b5}: '⦵'

CircledVerticalBar

\u{29b6}: '⦶'

CircledParallel

\u{29b7}: '⦷'

CircledReverseSolidus

\u{29b8}: '⦸'

CircledPerpendicular

\u{29b9}: '⦹'

CircleDividedByHorizontalBarAndTopHalfDividedByVerticalBar

\u{29ba}: '⦺'

CircleWithSuperimposedX

\u{29bb}: '⦻'

CircledAnticlockwiseDashRotatedDivisionSign

\u{29bc}: '⦼'

UpArrowThroughCircle

\u{29bd}: '⦽'

CircledWhiteBullet

\u{29be}: '⦾'

CircledBullet

\u{29bf}: '⦿'

CircledLessDashThan

\u{29c0}: '⧀'

CircledGreaterDashThan

\u{29c1}: '⧁'

CircleWithSmallCircleToTheRight

\u{29c2}: '⧂'

CircleWithTwoHorizontalStrokesToTheRight

\u{29c3}: '⧃'

SquaredRisingDiagonalSlash

\u{29c4}: '⧄'

SquaredFallingDiagonalSlash

\u{29c5}: '⧅'

SquaredAsterisk

\u{29c6}: '⧆'

SquaredSmallCircle

\u{29c7}: '⧇'

SquaredSquare

\u{29c8}: '⧈'

TwoJoinedSquares

\u{29c9}: '⧉'

TriangleWithDotAbove

\u{29ca}: '⧊'

TriangleWithUnderbar

\u{29cb}: '⧋'

SInTriangle

\u{29cc}: '⧌'

TriangleWithSerifsAtBottom

\u{29cd}: '⧍'

RightTriangleAboveLeftTriangle

\u{29ce}: '⧎'

LeftTriangleBesideVerticalBar

\u{29cf}: '⧏'

VerticalBarBesideRightTriangle

\u{29d0}: '⧐'

BowtieWithLeftHalfBlack

\u{29d1}: '⧑'

BowtieWithRightHalfBlack

\u{29d2}: '⧒'

BlackBowtie

\u{29d3}: '⧓'

TimesWithLeftHalfBlack

\u{29d4}: '⧔'

TimesWithRightHalfBlack

\u{29d5}: '⧕'

WhiteHourglass

\u{29d6}: '⧖'

BlackHourglass

\u{29d7}: '⧗'

LeftWigglyFence

\u{29d8}: '⧘'

RightWigglyFence

\u{29d9}: '⧙'

LeftDoubleWigglyFence

\u{29da}: '⧚'

RightDoubleWigglyFence

\u{29db}: '⧛'

IncompleteInfinity

\u{29dc}: '⧜'

TieOverInfinity

\u{29dd}: '⧝'

InfinityNegatedWithVerticalBar

\u{29de}: '⧞'

DoubleDashEndedMultimap

\u{29df}: '⧟'

SquareWithContouredOutline

\u{29e0}: '⧠'

IncreasesAs

\u{29e1}: '⧡'

ShuffleProduct

\u{29e2}: '⧢'

EqualsSignAndSlantedParallel

\u{29e3}: '⧣'

EqualsSignAndSlantedParallelWithTildeAbove

\u{29e4}: '⧤'

IdenticalToAndSlantedParallel

\u{29e5}: '⧥'

GleichStark

\u{29e6}: '⧦'

Thermodynamic

\u{29e7}: '⧧'

DownDashPointingTriangleWithLeftHalfBlack

\u{29e8}: '⧨'

DownDashPointingTriangleWithRightHalfBlack

\u{29e9}: '⧩'

BlackDiamondWithDownArrow

\u{29ea}: '⧪'

BlackLozenge

\u{29eb}: '⧫'

WhiteCircleWithDownArrow

\u{29ec}: '⧬'

BlackCircleWithDownArrow

\u{29ed}: '⧭'

ErrorDashBarredWhiteSquare

\u{29ee}: '⧮'

ErrorDashBarredBlackSquare

\u{29ef}: '⧯'

ErrorDashBarredWhiteDiamond

\u{29f0}: '⧰'

ErrorDashBarredBlackDiamond

\u{29f1}: '⧱'

ErrorDashBarredWhiteCircle

\u{29f2}: '⧲'

ErrorDashBarredBlackCircle

\u{29f3}: '⧳'

RuleDashDelayed

\u{29f4}: '⧴'

ReverseSolidusOperator

\u{29f5}: '⧵'

SolidusWithOverbar

\u{29f6}: '⧶'

ReverseSolidusWithHorizontalStroke

\u{29f7}: '⧷'

BigSolidus

\u{29f8}: '⧸'

BigReverseSolidus

\u{29f9}: '⧹'

DoublePlus

\u{29fa}: '⧺'

TriplePlus

\u{29fb}: '⧻'

LeftDashPointingCurvedAngleBracket

\u{29fc}: '⧼'

RightDashPointingCurvedAngleBracket

\u{29fd}: '⧽'

Tiny

\u{29fe}: '⧾'

Methods

impl MiscellaneousMathematicalSymbolsB

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case