Enum unicode_types::miscellaneous_symbols::MiscellaneousSymbols

pub enum MiscellaneousSymbols {
    BlackSunWithRays,
    Cloud,
    Umbrella,
    Snowman,
    Comet,
    BlackStar,
    WhiteStar,
    Lightning,
    Thunderstorm,
    Sun,
    AscendingNode,
    DescendingNode,
    Conjunction,
    Opposition,
    BlackTelephone,
    WhiteTelephone,
    BallotBox,
    BallotBoxWithCheck,
    BallotBoxWithX,
    Saltire,
    UmbrellaWithRainDrops,
    HotBeverage,
    WhiteShogiPiece,
    BlackShogiPiece,
    Shamrock,
    ReversedRotatedFloralHeartBullet,
    BlackLeftPointingIndex,
    BlackRightPointingIndex,
    WhiteLeftPointingIndex,
    WhiteUpPointingIndex,
    WhiteRightPointingIndex,
    WhiteDownPointingIndex,
    SkullAndCrossbones,
    CautionSign,
    RadioactiveSign,
    BiohazardSign,
    Caduceus,
    Ankh,
    OrthodoxCross,
    ChiRho,
    CrossOfLorraine,
    CrossOfJerusalem,
    StarAndCrescent,
    FarsiSymbol,
    AdiShakti,
    HammerAndSickle,
    PeaceSymbol,
    YinYang,
    TrigramForHeaven,
    TrigramForLake,
    TrigramForFire,
    TrigramForThunder,
    TrigramForWind,
    TrigramForWater,
    TrigramForMountain,
    TrigramForEarth,
    WheelOfDharma,
    WhiteFrowningFace,
    WhiteSmilingFace,
    BlackSmilingFace,
    WhiteSunWithRays,
    FirstQuarterMoon,
    LastQuarterMoon,
    Mercury,
    FemaleSign,
    Earth,
    MaleSign,
    Jupiter,
    Saturn,
    Uranus,
    Neptune,
    Pluto,
    Aries,
    Taurus,
    Gemini,
    Cancer,
    Leo,
    Virgo,
    Libra,
    Scorpius,
    Sagittarius,
    Capricorn,
    Aquarius,
    Pisces,
    WhiteChessKing,
    WhiteChessQueen,
    WhiteChessRook,
    WhiteChessBishop,
    WhiteChessKnight,
    WhiteChessPawn,
    BlackChessKing,
    BlackChessQueen,
    BlackChessRook,
    BlackChessBishop,
    BlackChessKnight,
    BlackChessPawn,
    BlackSpadeSuit,
    WhiteHeartSuit,
    WhiteDiamondSuit,
    BlackClubSuit,
    WhiteSpadeSuit,
    BlackHeartSuit,
    BlackDiamondSuit,
    WhiteClubSuit,
    HotSprings,
    QuarterNote,
    EighthNote,
    BeamedEighthNotes,
    BeamedSixteenthNotes,
    MusicFlatSign,
    MusicNaturalSign,
    MusicSharpSign,
    WestSyriacCross,
    EastSyriacCross,
    UniversalRecyclingSymbol,
    RecyclingSymbolForTypeDash1Plastics,
    RecyclingSymbolForTypeDash2Plastics,
    RecyclingSymbolForTypeDash3Plastics,
    RecyclingSymbolForTypeDash4Plastics,
    RecyclingSymbolForTypeDash5Plastics,
    RecyclingSymbolForTypeDash6Plastics,
    RecyclingSymbolForTypeDash7Plastics,
    RecyclingSymbolForGenericMaterials,
    BlackUniversalRecyclingSymbol,
    RecycledPaperSymbol,
    PartiallyDashRecycledPaperSymbol,
    PermanentPaperSign,
    WheelchairSymbol,
    DieFaceDash1,
    DieFaceDash2,
    DieFaceDash3,
    DieFaceDash4,
    DieFaceDash5,
    DieFaceDash6,
    WhiteCircleWithDotRight,
    WhiteCircleWithTwoDots,
    BlackCircleWithWhiteDotRight,
    BlackCircleWithTwoWhiteDots,
    MonogramForYang,
    MonogramForYin,
    DigramForGreaterYang,
    DigramForLesserYin,
    DigramForLesserYang,
    DigramForGreaterYin,
    WhiteFlag,
    BlackFlag,
    HammerAndPick,
    Anchor,
    CrossedSwords,
    StaffOfAesculapius,
    Scales,
    Alembic,
    Flower,
    Gear,
    StaffOfHermes,
    AtomSymbol,
    FleurDashDeDashLis,
    OutlinedWhiteStar,
    ThreeLinesConvergingRight,
    ThreeLinesConvergingLeft,
    WarningSign,
    HighVoltageSign,
    DoubledFemaleSign,
    DoubledMaleSign,
    InterlockedFemaleAndMaleSign,
    MaleAndFemaleSign,
    MaleWithStrokeSign,
    MaleWithStrokeAndMaleAndFemaleSign,
    VerticalMaleWithStrokeSign,
    HorizontalMaleWithStrokeSign,
    MediumWhiteCircle,
    MediumBlackCircle,
    MediumSmallWhiteCircle,
    MarriageSymbol,
    DivorceSymbol,
    UnmarriedPartnershipSymbol,
    Coffin,
    FuneralUrn,
    Neuter,
    Ceres,
    Pallas,
    Juno,
    Vesta,
    Chiron,
    BlackMoonLilith,
    Sextile,
    Semisextile,
    Quincunx,
    Sesquiquadrate,
    SoccerBall,
    Baseball,
    SquaredKey,
    WhiteDraughtsMan,
    WhiteDraughtsKing,
    BlackDraughtsMan,
    BlackDraughtsKing,
    SnowmanWithoutSnow,
    SunBehindCloud,
    Rain,
    BlackSnowman,
    ThunderCloudAndRain,
    TurnedWhiteShogiPiece,
    TurnedBlackShogiPiece,
    WhiteDiamondInSquare,
    CrossingLanes,
    DisabledCar,
    Ophiuchus,
    Pick,
    CarSliding,
    HelmetWithWhiteCross,
    CircledCrossingLanes,
    Chains,
    NoEntry,
    AlternateOneDashWayLeftWayTraffic,
    BlackTwoDashWayLeftWayTraffic,
    WhiteTwoDashWayLeftWayTraffic,
    BlackLeftLaneMerge,
    WhiteLeftLaneMerge,
    DriveSlowSign,
    HeavyWhiteDownDashPointingTriangle,
    LeftClosedEntry,
    SquaredSaltire,
    FallingDiagonalInWhiteCircleInBlackSquare,
    BlackTruck,
    RestrictedLeftEntryDash1,
    RestrictedLeftEntryDash2,
    AstronomicalSymbolForUranus,
    HeavyCircleWithStrokeAndTwoDotsAbove,
    Pentagram,
    RightDashHandedInterlacedPentagram,
    LeftDashHandedInterlacedPentagram,
    InvertedPentagram,
    BlackCrossOnShield,
    ShintoShrine,
    Church,
    Castle,
    HistoricSite,
    GearWithoutHub,
    GearWithHandles,
    MapSymbolForLighthouse,
    Mountain,
    UmbrellaOnGround,
    Fountain,
    FlagInHole,
    Ferry,
    Sailboat,
    SquareFourCorners,
    Skier,
    IceSkate,
    PersonWithBall,
    Tent,
    JapaneseBankSymbol,
    HeadstoneGraveyardSymbol,
    FuelPump,
    CupOnBlackSquare,
}

An enum to represent all characters in the MiscellaneousSymbols block.

Variants
BlackSunWithRays

\u{2600}: '☀'

Cloud

\u{2601}: '☁'

Umbrella

\u{2602}: '☂'

Snowman

\u{2603}: '☃'

Comet

\u{2604}: '☄'

BlackStar

\u{2605}: '★'

WhiteStar

\u{2606}: '☆'

Lightning

\u{2607}: '☇'

Thunderstorm

\u{2608}: '☈'

Sun

\u{2609}: '☉'

AscendingNode

\u{260a}: '☊'

DescendingNode

\u{260b}: '☋'

Conjunction

\u{260c}: '☌'

Opposition

\u{260d}: '☍'

BlackTelephone

\u{260e}: '☎'

WhiteTelephone

\u{260f}: '☏'

BallotBox

\u{2610}: '☐'

BallotBoxWithCheck

\u{2611}: '☑'

BallotBoxWithX

\u{2612}: '☒'

Saltire

\u{2613}: '☓'

UmbrellaWithRainDrops

\u{2614}: '☔'

HotBeverage

\u{2615}: '☕'

WhiteShogiPiece

\u{2616}: '☖'

BlackShogiPiece

\u{2617}: '☗'

Shamrock

\u{2618}: '☘'

ReversedRotatedFloralHeartBullet

\u{2619}: '☙'

BlackLeftPointingIndex

\u{261a}: '☚'

BlackRightPointingIndex

\u{261b}: '☛'

WhiteLeftPointingIndex

\u{261c}: '☜'

WhiteUpPointingIndex

\u{261d}: '☝'

WhiteRightPointingIndex

\u{261e}: '☞'

WhiteDownPointingIndex

\u{261f}: '☟'

SkullAndCrossbones

\u{2620}: '☠'

CautionSign

\u{2621}: '☡'

RadioactiveSign

\u{2622}: '☢'

BiohazardSign

\u{2623}: '☣'

Caduceus

\u{2624}: '☤'

Ankh

\u{2625}: '☥'

OrthodoxCross

\u{2626}: '☦'

ChiRho

\u{2627}: '☧'

CrossOfLorraine

\u{2628}: '☨'

CrossOfJerusalem

\u{2629}: '☩'

StarAndCrescent

\u{262a}: '☪'

FarsiSymbol

\u{262b}: '☫'

AdiShakti

\u{262c}: '☬'

HammerAndSickle

\u{262d}: '☭'

PeaceSymbol

\u{262e}: '☮'

YinYang

\u{262f}: '☯'

TrigramForHeaven

\u{2630}: '☰'

TrigramForLake

\u{2631}: '☱'

TrigramForFire

\u{2632}: '☲'

TrigramForThunder

\u{2633}: '☳'

TrigramForWind

\u{2634}: '☴'

TrigramForWater

\u{2635}: '☵'

TrigramForMountain

\u{2636}: '☶'

TrigramForEarth

\u{2637}: '☷'

WheelOfDharma

\u{2638}: '☸'

WhiteFrowningFace

\u{2639}: '☹'

WhiteSmilingFace

\u{263a}: '☺'

BlackSmilingFace

\u{263b}: '☻'

WhiteSunWithRays

\u{263c}: '☼'

FirstQuarterMoon

\u{263d}: '☽'

LastQuarterMoon

\u{263e}: '☾'

Mercury

\u{263f}: '☿'

FemaleSign

\u{2640}: '♀'

Earth

\u{2641}: '♁'

MaleSign

\u{2642}: '♂'

Jupiter

\u{2643}: '♃'

Saturn

\u{2644}: '♄'

Uranus

\u{2645}: '♅'

Neptune

\u{2646}: '♆'

Pluto

\u{2647}: '♇'

Aries

\u{2648}: '♈'

Taurus

\u{2649}: '♉'

Gemini

\u{264a}: '♊'

Cancer

\u{264b}: '♋'

Leo

\u{264c}: '♌'

Virgo

\u{264d}: '♍'

Libra

\u{264e}: '♎'

Scorpius

\u{264f}: '♏'

Sagittarius

\u{2650}: '♐'

Capricorn

\u{2651}: '♑'

Aquarius

\u{2652}: '♒'

Pisces

\u{2653}: '♓'

WhiteChessKing

\u{2654}: '♔'

WhiteChessQueen

\u{2655}: '♕'

WhiteChessRook

\u{2656}: '♖'

WhiteChessBishop

\u{2657}: '♗'

WhiteChessKnight

\u{2658}: '♘'

WhiteChessPawn

\u{2659}: '♙'

BlackChessKing

\u{265a}: '♚'

BlackChessQueen

\u{265b}: '♛'

BlackChessRook

\u{265c}: '♜'

BlackChessBishop

\u{265d}: '♝'

BlackChessKnight

\u{265e}: '♞'

BlackChessPawn

\u{265f}: '♟'

BlackSpadeSuit

\u{2660}: '♠'

WhiteHeartSuit

\u{2661}: '♡'

WhiteDiamondSuit

\u{2662}: '♢'

BlackClubSuit

\u{2663}: '♣'

WhiteSpadeSuit

\u{2664}: '♤'

BlackHeartSuit

\u{2665}: '♥'

BlackDiamondSuit

\u{2666}: '♦'

WhiteClubSuit

\u{2667}: '♧'

HotSprings

\u{2668}: '♨'

QuarterNote

\u{2669}: '♩'

EighthNote

\u{266a}: '♪'

BeamedEighthNotes

\u{266b}: '♫'

BeamedSixteenthNotes

\u{266c}: '♬'

MusicFlatSign

\u{266d}: '♭'

MusicNaturalSign

\u{266e}: '♮'

MusicSharpSign

\u{266f}: '♯'

WestSyriacCross

\u{2670}: '♰'

EastSyriacCross

\u{2671}: '♱'

UniversalRecyclingSymbol

\u{2672}: '♲'

RecyclingSymbolForTypeDash1Plastics

\u{2673}: '♳'

RecyclingSymbolForTypeDash2Plastics

\u{2674}: '♴'

RecyclingSymbolForTypeDash3Plastics

\u{2675}: '♵'

RecyclingSymbolForTypeDash4Plastics

\u{2676}: '♶'

RecyclingSymbolForTypeDash5Plastics

\u{2677}: '♷'

RecyclingSymbolForTypeDash6Plastics

\u{2678}: '♸'

RecyclingSymbolForTypeDash7Plastics

\u{2679}: '♹'

RecyclingSymbolForGenericMaterials

\u{267a}: '♺'

BlackUniversalRecyclingSymbol

\u{267b}: '♻'

RecycledPaperSymbol

\u{267c}: '♼'

PartiallyDashRecycledPaperSymbol

\u{267d}: '♽'

PermanentPaperSign

\u{267e}: '♾'

WheelchairSymbol

\u{267f}: '♿'

DieFaceDash1

\u{2680}: '⚀'

DieFaceDash2

\u{2681}: '⚁'

DieFaceDash3

\u{2682}: '⚂'

DieFaceDash4

\u{2683}: '⚃'

DieFaceDash5

\u{2684}: '⚄'

DieFaceDash6

\u{2685}: '⚅'

WhiteCircleWithDotRight

\u{2686}: '⚆'

WhiteCircleWithTwoDots

\u{2687}: '⚇'

BlackCircleWithWhiteDotRight

\u{2688}: '⚈'

BlackCircleWithTwoWhiteDots

\u{2689}: '⚉'

MonogramForYang

\u{268a}: '⚊'

MonogramForYin

\u{268b}: '⚋'

DigramForGreaterYang

\u{268c}: '⚌'

DigramForLesserYin

\u{268d}: '⚍'

DigramForLesserYang

\u{268e}: '⚎'

DigramForGreaterYin

\u{268f}: '⚏'

WhiteFlag

\u{2690}: '⚐'

BlackFlag

\u{2691}: '⚑'

HammerAndPick

\u{2692}: '⚒'

Anchor

\u{2693}: '⚓'

CrossedSwords

\u{2694}: '⚔'

StaffOfAesculapius

\u{2695}: '⚕'

Scales

\u{2696}: '⚖'

Alembic

\u{2697}: '⚗'

Flower

\u{2698}: '⚘'

Gear

\u{2699}: '⚙'

StaffOfHermes

\u{269a}: '⚚'

AtomSymbol

\u{269b}: '⚛'

FleurDashDeDashLis

\u{269c}: '⚜'

OutlinedWhiteStar

\u{269d}: '⚝'

ThreeLinesConvergingRight

\u{269e}: '⚞'

ThreeLinesConvergingLeft

\u{269f}: '⚟'

WarningSign

\u{26a0}: '⚠'

HighVoltageSign

\u{26a1}: '⚡'

DoubledFemaleSign

\u{26a2}: '⚢'

DoubledMaleSign

\u{26a3}: '⚣'

InterlockedFemaleAndMaleSign

\u{26a4}: '⚤'

MaleAndFemaleSign

\u{26a5}: '⚥'

MaleWithStrokeSign

\u{26a6}: '⚦'

MaleWithStrokeAndMaleAndFemaleSign

\u{26a7}: '⚧'

VerticalMaleWithStrokeSign

\u{26a8}: '⚨'

HorizontalMaleWithStrokeSign

\u{26a9}: '⚩'

MediumWhiteCircle

\u{26aa}: '⚪'

MediumBlackCircle

\u{26ab}: '⚫'

MediumSmallWhiteCircle

\u{26ac}: '⚬'

MarriageSymbol

\u{26ad}: '⚭'

DivorceSymbol

\u{26ae}: '⚮'

UnmarriedPartnershipSymbol

\u{26af}: '⚯'

Coffin

\u{26b0}: '⚰'

FuneralUrn

\u{26b1}: '⚱'

Neuter

\u{26b2}: '⚲'

Ceres

\u{26b3}: '⚳'

Pallas

\u{26b4}: '⚴'

Juno

\u{26b5}: '⚵'

Vesta

\u{26b6}: '⚶'

Chiron

\u{26b7}: '⚷'

BlackMoonLilith

\u{26b8}: '⚸'

Sextile

\u{26b9}: '⚹'

Semisextile

\u{26ba}: '⚺'

Quincunx

\u{26bb}: '⚻'

Sesquiquadrate

\u{26bc}: '⚼'

SoccerBall

\u{26bd}: '⚽'

Baseball

\u{26be}: '⚾'

SquaredKey

\u{26bf}: '⚿'

WhiteDraughtsMan

\u{26c0}: '⛀'

WhiteDraughtsKing

\u{26c1}: '⛁'

BlackDraughtsMan

\u{26c2}: '⛂'

BlackDraughtsKing

\u{26c3}: '⛃'

SnowmanWithoutSnow

\u{26c4}: '⛄'

SunBehindCloud

\u{26c5}: '⛅'

Rain

\u{26c6}: '⛆'

BlackSnowman

\u{26c7}: '⛇'

ThunderCloudAndRain

\u{26c8}: '⛈'

TurnedWhiteShogiPiece

\u{26c9}: '⛉'

TurnedBlackShogiPiece

\u{26ca}: '⛊'

WhiteDiamondInSquare

\u{26cb}: '⛋'

CrossingLanes

\u{26cc}: '⛌'

DisabledCar

\u{26cd}: '⛍'

Ophiuchus

\u{26ce}: '⛎'

Pick

\u{26cf}: '⛏'

CarSliding

\u{26d0}: '⛐'

HelmetWithWhiteCross

\u{26d1}: '⛑'

CircledCrossingLanes

\u{26d2}: '⛒'

Chains

\u{26d3}: '⛓'

NoEntry

\u{26d4}: '⛔'

AlternateOneDashWayLeftWayTraffic

\u{26d5}: '⛕'

BlackTwoDashWayLeftWayTraffic

\u{26d6}: '⛖'

WhiteTwoDashWayLeftWayTraffic

\u{26d7}: '⛗'

BlackLeftLaneMerge

\u{26d8}: '⛘'

WhiteLeftLaneMerge

\u{26d9}: '⛙'

DriveSlowSign

\u{26da}: '⛚'

HeavyWhiteDownDashPointingTriangle

\u{26db}: '⛛'

LeftClosedEntry

\u{26dc}: '⛜'

SquaredSaltire

\u{26dd}: '⛝'

FallingDiagonalInWhiteCircleInBlackSquare

\u{26de}: '⛞'

BlackTruck

\u{26df}: '⛟'

RestrictedLeftEntryDash1

\u{26e0}: '⛠'

RestrictedLeftEntryDash2

\u{26e1}: '⛡'

AstronomicalSymbolForUranus

\u{26e2}: '⛢'

HeavyCircleWithStrokeAndTwoDotsAbove

\u{26e3}: '⛣'

Pentagram

\u{26e4}: '⛤'

RightDashHandedInterlacedPentagram

\u{26e5}: '⛥'

LeftDashHandedInterlacedPentagram

\u{26e6}: '⛦'

InvertedPentagram

\u{26e7}: '⛧'

BlackCrossOnShield

\u{26e8}: '⛨'

ShintoShrine

\u{26e9}: '⛩'

Church

\u{26ea}: '⛪'

Castle

\u{26eb}: '⛫'

HistoricSite

\u{26ec}: '⛬'

GearWithoutHub

\u{26ed}: '⛭'

GearWithHandles

\u{26ee}: '⛮'

MapSymbolForLighthouse

\u{26ef}: '⛯'

Mountain

\u{26f0}: '⛰'

UmbrellaOnGround

\u{26f1}: '⛱'

Fountain

\u{26f2}: '⛲'

FlagInHole

\u{26f3}: '⛳'

Ferry

\u{26f4}: '⛴'

Sailboat

\u{26f5}: '⛵'

SquareFourCorners

\u{26f6}: '⛶'

Skier

\u{26f7}: '⛷'

IceSkate

\u{26f8}: '⛸'

PersonWithBall

\u{26f9}: '⛹'

Tent

\u{26fa}: '⛺'

JapaneseBankSymbol

\u{26fb}: '⛻'

HeadstoneGraveyardSymbol

\u{26fc}: '⛼'

FuelPump

\u{26fd}: '⛽'

CupOnBlackSquare

\u{26fe}: '⛾'

Methods

impl MiscellaneousSymbols

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case