Enum unicode_types::miscellaneous_symbols_and_arrows::MiscellaneousSymbolsandArrows

pub enum MiscellaneousSymbolsandArrows {
    NorthEastWhiteArrow,
    NorthWestWhiteArrow,
    SouthEastWhiteArrow,
    SouthWestWhiteArrow,
    LeftRightWhiteArrow,
    LeftwardsBlackArrow,
    UpwardsBlackArrow,
    DownwardsBlackArrow,
    NorthEastBlackArrow,
    NorthWestBlackArrow,
    SouthEastBlackArrow,
    SouthWestBlackArrow,
    LeftRightBlackArrow,
    UpDownBlackArrow,
    RightwardsArrowWithTipDownwards,
    RightwardsArrowWithTipUpwards,
    LeftwardsArrowWithTipDownwards,
    LeftwardsArrowWithTipUpwards,
    SquareWithTopHalfBlack,
    SquareWithBottomHalfBlack,
    SquareWithUpperRightDiagonalHalfBlack,
    SquareWithLowerLeftDiagonalHalfBlack,
    DiamondWithLeftHalfBlack,
    DiamondWithRightHalfBlack,
    DiamondWithTopHalfBlack,
    DiamondWithBottomHalfBlack,
    DottedSquare,
    BlackLargeSquare,
    WhiteLargeSquare,
    BlackVerySmallSquare,
    WhiteVerySmallSquare,
    BlackPentagon,
    WhitePentagon,
    WhiteHexagon,
    BlackHexagon,
    HorizontalBlackHexagon,
    BlackLargeCircle,
    BlackMediumDiamond,
    WhiteMediumDiamond,
    BlackMediumLozenge,
    WhiteMediumLozenge,
    BlackSmallDiamond,
    BlackSmallLozenge,
    WhiteSmallLozenge,
    BlackHorizontalEllipse,
    WhiteHorizontalEllipse,
    BlackVerticalEllipse,
    WhiteVerticalEllipse,
    LeftArrowWithSmallCircle,
    ThreeLeftwardsArrows,
    LeftArrowWithCircledPlus,
    LongLeftwardsSquiggleArrow,
    LeftwardsTwoDashHeadedArrowWithVerticalStroke,
    LeftwardsTwoDashHeadedArrowWithDoubleVerticalStroke,
    LeftwardsTwoDashHeadedArrowFromBar,
    LeftwardsTwoDashHeadedTripleDashArrow,
    LeftwardsArrowWithDottedStem,
    LeftwardsArrowWithTailWithVerticalStroke,
    LeftwardsArrowWithTailWithDoubleVerticalStroke,
    LeftwardsTwoDashHeadedArrowWithTail,
    LeftwardsTwoDashHeadedArrowWithTailWithVerticalStroke,
    LeftwardsTwoDashHeadedArrowWithTailWithDoubleVerticalStroke,
    LeftwardsArrowThroughX,
    WaveArrowPointingDirectlyLeft,
    EqualsSignAboveLeftwardsArrow,
    ReverseTildeOperatorAboveLeftwardsArrow,
    LeftwardsArrowAboveReverseAlmostEqualTo,
    RightwardsArrowThroughGreaterDashThan,
    RightwardsArrowThroughSuperset,
    LeftwardsQuadrupleArrow,
    RightwardsQuadrupleArrow,
    ReverseTildeOperatorAboveRightwardsArrow,
    RightwardsArrowAboveReverseAlmostEqualTo,
    TildeOperatorAboveLeftwardsArrow,
    LeftwardsArrowAboveAlmostEqualTo,
    LeftwardsArrowAboveReverseTildeOperator,
    RightwardsArrowAboveReverseTildeOperator,
    DownwardsTriangleDashHeadedZigzagArrow,
    ShortSlantedNorthArrow,
    ShortBackslantedSouthArrow,
    WhiteMediumStar,
    BlackSmallStar,
    WhiteSmallStar,
    BlackRightDashPointingPentagon,
    WhiteRightDashPointingPentagon,
    HeavyLargeCircle,
    HeavyOvalWithOvalInside,
    HeavyCircleWithCircleInside,
    HeavyCircle,
    HeavyCircledSaltire,
    SlantedNorthArrowWithHookedHead,
    BackslantedSouthArrowWithHookedTail,
    SlantedNorthArrowWithHorizontalTail,
    BackslantedSouthArrowWithHorizontalTail,
    BentArrowPointingDownwardsThenNorthEast,
    ShortBentArrowPointingDownwardsThenNorthEast,
    LeftwardsTriangleDashHeadedArrow,
    UpwardsTriangleDashHeadedArrow,
    RightwardsTriangleDashHeadedArrow,
    DownwardsTriangleDashHeadedArrow,
    LeftRightTriangleDashHeadedArrow,
    UpDownTriangleDashHeadedArrow,
    NorthWestTriangleDashHeadedArrow,
    NorthEastTriangleDashHeadedArrow,
    SouthEastTriangleDashHeadedArrow,
    SouthWestTriangleDashHeadedArrow,
    LeftwardsTriangleDashHeadedDashedArrow,
    UpwardsTriangleDashHeadedDashedArrow,
    RightwardsTriangleDashHeadedDashedArrow,
    DownwardsTriangleDashHeadedDashedArrow,
    ClockwiseTriangleDashHeadedOpenCircleArrow,
    AnticlockwiseTriangleDashHeadedOpenCircleArrow,
    LeftwardsTriangleDashHeadedArrowToBar,
    UpwardsTriangleDashHeadedArrowToBar,
    RightwardsTriangleDashHeadedArrowToBar,
    DownwardsTriangleDashHeadedArrowToBar,
    NorthWestTriangleDashHeadedArrowToBar,
    NorthEastTriangleDashHeadedArrowToBar,
    SouthEastTriangleDashHeadedArrowToBar,
    SouthWestTriangleDashHeadedArrowToBar,
    LeftwardsTriangleDashHeadedArrowWithDoubleHorizontalStroke,
    UpwardsTriangleDashHeadedArrowWithDoubleHorizontalStroke,
    RightwardsTriangleDashHeadedArrowWithDoubleHorizontalStroke,
    DownwardsTriangleDashHeadedArrowWithDoubleHorizontalStroke,
    HorizontalTabKey,
    VerticalTabKey,
    LeftwardsTriangleDashHeadedArrowOverRightwardsTriangleDashHeadedArrow,
    UpwardsTriangleDashHeadedArrowLeftwardsOfDownwardsTriangleDashHeadedArrow,
    RightwardsTriangleDashHeadedArrowOverLeftwardsTriangleDashHeadedArrow,
    DownwardsTriangleDashHeadedArrowLeftwardsOfUpwardsTriangleDashHeadedArrow,
    LeftwardsTriangleDashHeadedPairedArrows,
    UpwardsTriangleDashHeadedPairedArrows,
    RightwardsTriangleDashHeadedPairedArrows,
    DownwardsTriangleDashHeadedPairedArrows,
    LeftwardsBlackCircledWhiteArrow,
    UpwardsBlackCircledWhiteArrow,
    RightwardsBlackCircledWhiteArrow,
    DownwardsBlackCircledWhiteArrow,
    AnticlockwiseTriangleDashHeadedRightUDashShapedArrow,
    AnticlockwiseTriangleDashHeadedBottomUDashShapedArrow,
    AnticlockwiseTriangleDashHeadedLeftUDashShapedArrow,
    AnticlockwiseTriangleDashHeadedTopUDashShapedArrow,
    ReturnLeft,
    ReturnRight,
    NewlineLeft,
    NewlineRight,
    FourCornerArrowsCirclingAnticlockwise,
    RightwardsBlackArrow,
    ThreeDashDTopDashLightedLeftwardsEquilateralArrowhead,
    ThreeDashDRightDashLightedUpwardsEquilateralArrowhead,
    ThreeDashDTopDashLightedRightwardsEquilateralArrowhead,
    ThreeDashDLeftDashLightedDownwardsEquilateralArrowhead,
    BlackLeftwardsEquilateralArrowhead,
    BlackUpwardsEquilateralArrowhead,
    BlackRightwardsEquilateralArrowhead,
    BlackDownwardsEquilateralArrowhead,
    DownwardsTriangleDashHeadedArrowWithLongTipLeftwards,
    DownwardsTriangleDashHeadedArrowWithLongTipRightwards,
    UpwardsTriangleDashHeadedArrowWithLongTipLeftwards,
    UpwardsTriangleDashHeadedArrowWithLongTipRightwards,
    LeftwardsTriangleDashHeadedArrowWithLongTipUpwards,
    RightwardsTriangleDashHeadedArrowWithLongTipUpwards,
    LeftwardsTriangleDashHeadedArrowWithLongTipDownwards,
    RightwardsTriangleDashHeadedArrowWithLongTipDownwards,
    BlackCurvedDownwardsAndLeftwardsArrow,
    BlackCurvedDownwardsAndRightwardsArrow,
    BlackCurvedUpwardsAndLeftwardsArrow,
    BlackCurvedUpwardsAndRightwardsArrow,
    BlackCurvedLeftwardsAndUpwardsArrow,
    BlackCurvedRightwardsAndUpwardsArrow,
    BlackCurvedLeftwardsAndDownwardsArrow,
    BlackCurvedRightwardsAndDownwardsArrow,
    RibbonArrowDownLeft,
    RibbonArrowDownRight,
    RibbonArrowUpLeft,
    RibbonArrowUpRight,
    RibbonArrowLeftUp,
    RibbonArrowRightUp,
    RibbonArrowLeftDown,
    RibbonArrowRightDown,
    UpwardsWhiteArrowFromBarWithHorizontalBar,
    UpArrowheadInARectangleBox,
    OverlappingWhiteSquares,
    OverlappingWhiteAndBlackSquares,
    OverlappingBlackSquares,
    BallotBoxWithLightX,
    CircledX,
    CircledBoldX,
    BlackSquareCentred,
    BlackDiamondCentred,
    TurnedBlackPentagon,
    HorizontalBlackOctagon,
    BlackOctagon,
    BlackMediumUpDashPointingTriangleCentred,
    BlackMediumDownDashPointingTriangleCentred,
    BlackMediumLeftDashPointingTriangleCentred,
    BlackMediumRightDashPointingTriangleCentred,
    NeptuneFormTwo,
    TopHalfBlackCircle,
    BottomHalfBlackCircle,
    LightFourPointedBlackCusp,
    RotatedLightFourPointedBlackCusp,
    WhiteFourPointedCusp,
    RotatedWhiteFourPointedCusp,
    SquarePositionIndicator,
    UncertaintySign,
    GroupMark,
    PlutoFormTwo,
    PlutoFormThree,
    PlutoFormFour,
    PlutoFormFive,
    Transpluto,
    Proserpina,
    Astraea,
    Hygiea,
    Pholus,
    Nessus,
    WhiteMoonSelena,
    BlackDiamondOnCross,
    TrueLightMoonArta,
    Cupido,
    Hades,
    Zeus,
    Kronos,
    Apollon,
    Admetos,
    Vulcanus,
    Poseidon,
    LeftHalfBlackStar,
    RightHalfBlackStar,
    StarWithLeftHalfBlack,
    StarWithRightHalfBlack,
    LeftwardsTwoDashHeadedArrowWithTriangleArrowheads,
    UpwardsTwoDashHeadedArrowWithTriangleArrowheads,
    RightwardsTwoDashHeadedArrowWithTriangleArrowheads,
    DownwardsTwoDashHeadedArrowWithTriangleArrowheads,
    ErisFormOne,
    ErisFormTwo,
    Sedna,
    RussianAstrologicalSymbolVigintile,
    RussianAstrologicalSymbolNovile,
    RussianAstrologicalSymbolQuintile,
    RussianAstrologicalSymbolBinovile,
    RussianAstrologicalSymbolSentagon,
    RussianAstrologicalSymbolTredecile,
    EqualsSignWithInfinityBelow,
    UnitedSymbol,
    SeparatedSymbol,
    DoubledSymbol,
    PassedSymbol,
    ReversedRightAngle,
}

An enum to represent all characters in the MiscellaneousSymbolsandArrows block.

Variants
NorthEastWhiteArrow

\u{2b00}: '⬀'

NorthWestWhiteArrow

\u{2b01}: '⬁'

SouthEastWhiteArrow

\u{2b02}: '⬂'

SouthWestWhiteArrow

\u{2b03}: '⬃'

LeftRightWhiteArrow

\u{2b04}: '⬄'

LeftwardsBlackArrow

\u{2b05}: '⬅'

UpwardsBlackArrow

\u{2b06}: '⬆'

DownwardsBlackArrow

\u{2b07}: '⬇'

NorthEastBlackArrow

\u{2b08}: '⬈'

NorthWestBlackArrow

\u{2b09}: '⬉'

SouthEastBlackArrow

\u{2b0a}: '⬊'

SouthWestBlackArrow

\u{2b0b}: '⬋'

LeftRightBlackArrow

\u{2b0c}: '⬌'

UpDownBlackArrow

\u{2b0d}: '⬍'

RightwardsArrowWithTipDownwards

\u{2b0e}: '⬎'

RightwardsArrowWithTipUpwards

\u{2b0f}: '⬏'

LeftwardsArrowWithTipDownwards

\u{2b10}: '⬐'

LeftwardsArrowWithTipUpwards

\u{2b11}: '⬑'

SquareWithTopHalfBlack

\u{2b12}: '⬒'

SquareWithBottomHalfBlack

\u{2b13}: '⬓'

SquareWithUpperRightDiagonalHalfBlack

\u{2b14}: '⬔'

SquareWithLowerLeftDiagonalHalfBlack

\u{2b15}: '⬕'

DiamondWithLeftHalfBlack

\u{2b16}: '⬖'

DiamondWithRightHalfBlack

\u{2b17}: '⬗'

DiamondWithTopHalfBlack

\u{2b18}: '⬘'

DiamondWithBottomHalfBlack

\u{2b19}: '⬙'

DottedSquare

\u{2b1a}: '⬚'

BlackLargeSquare

\u{2b1b}: '⬛'

WhiteLargeSquare

\u{2b1c}: '⬜'

BlackVerySmallSquare

\u{2b1d}: '⬝'

WhiteVerySmallSquare

\u{2b1e}: '⬞'

BlackPentagon

\u{2b1f}: '⬟'

WhitePentagon

\u{2b20}: '⬠'

WhiteHexagon

\u{2b21}: '⬡'

BlackHexagon

\u{2b22}: '⬢'

HorizontalBlackHexagon

\u{2b23}: '⬣'

BlackLargeCircle

\u{2b24}: '⬤'

BlackMediumDiamond

\u{2b25}: '⬥'

WhiteMediumDiamond

\u{2b26}: '⬦'

BlackMediumLozenge

\u{2b27}: '⬧'

WhiteMediumLozenge

\u{2b28}: '⬨'

BlackSmallDiamond

\u{2b29}: '⬩'

BlackSmallLozenge

\u{2b2a}: '⬪'

WhiteSmallLozenge

\u{2b2b}: '⬫'

BlackHorizontalEllipse

\u{2b2c}: '⬬'

WhiteHorizontalEllipse

\u{2b2d}: '⬭'

BlackVerticalEllipse

\u{2b2e}: '⬮'

WhiteVerticalEllipse

\u{2b2f}: '⬯'

LeftArrowWithSmallCircle

\u{2b30}: '⬰'

ThreeLeftwardsArrows

\u{2b31}: '⬱'

LeftArrowWithCircledPlus

\u{2b32}: '⬲'

LongLeftwardsSquiggleArrow

\u{2b33}: '⬳'

LeftwardsTwoDashHeadedArrowWithVerticalStroke

\u{2b34}: '⬴'

LeftwardsTwoDashHeadedArrowWithDoubleVerticalStroke

\u{2b35}: '⬵'

LeftwardsTwoDashHeadedArrowFromBar

\u{2b36}: '⬶'

LeftwardsTwoDashHeadedTripleDashArrow

\u{2b37}: '⬷'

LeftwardsArrowWithDottedStem

\u{2b38}: '⬸'

LeftwardsArrowWithTailWithVerticalStroke

\u{2b39}: '⬹'

LeftwardsArrowWithTailWithDoubleVerticalStroke

\u{2b3a}: '⬺'

LeftwardsTwoDashHeadedArrowWithTail

\u{2b3b}: '⬻'

LeftwardsTwoDashHeadedArrowWithTailWithVerticalStroke

\u{2b3c}: '⬼'

LeftwardsTwoDashHeadedArrowWithTailWithDoubleVerticalStroke

\u{2b3d}: '⬽'

LeftwardsArrowThroughX

\u{2b3e}: '⬾'

WaveArrowPointingDirectlyLeft

\u{2b3f}: '⬿'

EqualsSignAboveLeftwardsArrow

\u{2b40}: '⭀'

ReverseTildeOperatorAboveLeftwardsArrow

\u{2b41}: '⭁'

LeftwardsArrowAboveReverseAlmostEqualTo

\u{2b42}: '⭂'

RightwardsArrowThroughGreaterDashThan

\u{2b43}: '⭃'

RightwardsArrowThroughSuperset

\u{2b44}: '⭄'

LeftwardsQuadrupleArrow

\u{2b45}: '⭅'

RightwardsQuadrupleArrow

\u{2b46}: '⭆'

ReverseTildeOperatorAboveRightwardsArrow

\u{2b47}: '⭇'

RightwardsArrowAboveReverseAlmostEqualTo

\u{2b48}: '⭈'

TildeOperatorAboveLeftwardsArrow

\u{2b49}: '⭉'

LeftwardsArrowAboveAlmostEqualTo

\u{2b4a}: '⭊'

LeftwardsArrowAboveReverseTildeOperator

\u{2b4b}: '⭋'

RightwardsArrowAboveReverseTildeOperator

\u{2b4c}: '⭌'

DownwardsTriangleDashHeadedZigzagArrow

\u{2b4d}: '⭍'

ShortSlantedNorthArrow

\u{2b4e}: '⭎'

ShortBackslantedSouthArrow

\u{2b4f}: '⭏'

WhiteMediumStar

\u{2b50}: '⭐'

BlackSmallStar

\u{2b51}: '⭑'

WhiteSmallStar

\u{2b52}: '⭒'

BlackRightDashPointingPentagon

\u{2b53}: '⭓'

WhiteRightDashPointingPentagon

\u{2b54}: '⭔'

HeavyLargeCircle

\u{2b55}: '⭕'

HeavyOvalWithOvalInside

\u{2b56}: '⭖'

HeavyCircleWithCircleInside

\u{2b57}: '⭗'

HeavyCircle

\u{2b58}: '⭘'

HeavyCircledSaltire

\u{2b59}: '⭙'

SlantedNorthArrowWithHookedHead

\u{2b5a}: '⭚'

BackslantedSouthArrowWithHookedTail

\u{2b5b}: '⭛'

SlantedNorthArrowWithHorizontalTail

\u{2b5c}: '⭜'

BackslantedSouthArrowWithHorizontalTail

\u{2b5d}: '⭝'

BentArrowPointingDownwardsThenNorthEast

\u{2b5e}: '⭞'

ShortBentArrowPointingDownwardsThenNorthEast

\u{2b5f}: '⭟'

LeftwardsTriangleDashHeadedArrow

\u{2b60}: '⭠'

UpwardsTriangleDashHeadedArrow

\u{2b61}: '⭡'

RightwardsTriangleDashHeadedArrow

\u{2b62}: '⭢'

DownwardsTriangleDashHeadedArrow

\u{2b63}: '⭣'

LeftRightTriangleDashHeadedArrow

\u{2b64}: '⭤'

UpDownTriangleDashHeadedArrow

\u{2b65}: '⭥'

NorthWestTriangleDashHeadedArrow

\u{2b66}: '⭦'

NorthEastTriangleDashHeadedArrow

\u{2b67}: '⭧'

SouthEastTriangleDashHeadedArrow

\u{2b68}: '⭨'

SouthWestTriangleDashHeadedArrow

\u{2b69}: '⭩'

LeftwardsTriangleDashHeadedDashedArrow

\u{2b6a}: '⭪'

UpwardsTriangleDashHeadedDashedArrow

\u{2b6b}: '⭫'

RightwardsTriangleDashHeadedDashedArrow

\u{2b6c}: '⭬'

DownwardsTriangleDashHeadedDashedArrow

\u{2b6d}: '⭭'

ClockwiseTriangleDashHeadedOpenCircleArrow

\u{2b6e}: '⭮'

AnticlockwiseTriangleDashHeadedOpenCircleArrow

\u{2b6f}: '⭯'

LeftwardsTriangleDashHeadedArrowToBar

\u{2b70}: '⭰'

UpwardsTriangleDashHeadedArrowToBar

\u{2b71}: '⭱'

RightwardsTriangleDashHeadedArrowToBar

\u{2b72}: '⭲'

DownwardsTriangleDashHeadedArrowToBar

\u{2b73}: '⭳'

NorthWestTriangleDashHeadedArrowToBar

\u{2b76}: '⭶'

NorthEastTriangleDashHeadedArrowToBar

\u{2b77}: '⭷'

SouthEastTriangleDashHeadedArrowToBar

\u{2b78}: '⭸'

SouthWestTriangleDashHeadedArrowToBar

\u{2b79}: '⭹'

LeftwardsTriangleDashHeadedArrowWithDoubleHorizontalStroke

\u{2b7a}: '⭺'

UpwardsTriangleDashHeadedArrowWithDoubleHorizontalStroke

\u{2b7b}: '⭻'

RightwardsTriangleDashHeadedArrowWithDoubleHorizontalStroke

\u{2b7c}: '⭼'

DownwardsTriangleDashHeadedArrowWithDoubleHorizontalStroke

\u{2b7d}: '⭽'

HorizontalTabKey

\u{2b7e}: '⭾'

VerticalTabKey

\u{2b7f}: '⭿'

LeftwardsTriangleDashHeadedArrowOverRightwardsTriangleDashHeadedArrow

\u{2b80}: '⮀'

UpwardsTriangleDashHeadedArrowLeftwardsOfDownwardsTriangleDashHeadedArrow

\u{2b81}: '⮁'

RightwardsTriangleDashHeadedArrowOverLeftwardsTriangleDashHeadedArrow

\u{2b82}: '⮂'

DownwardsTriangleDashHeadedArrowLeftwardsOfUpwardsTriangleDashHeadedArrow

\u{2b83}: '⮃'

LeftwardsTriangleDashHeadedPairedArrows

\u{2b84}: '⮄'

UpwardsTriangleDashHeadedPairedArrows

\u{2b85}: '⮅'

RightwardsTriangleDashHeadedPairedArrows

\u{2b86}: '⮆'

DownwardsTriangleDashHeadedPairedArrows

\u{2b87}: '⮇'

LeftwardsBlackCircledWhiteArrow

\u{2b88}: '⮈'

UpwardsBlackCircledWhiteArrow

\u{2b89}: '⮉'

RightwardsBlackCircledWhiteArrow

\u{2b8a}: '⮊'

DownwardsBlackCircledWhiteArrow

\u{2b8b}: '⮋'

AnticlockwiseTriangleDashHeadedRightUDashShapedArrow

\u{2b8c}: '⮌'

AnticlockwiseTriangleDashHeadedBottomUDashShapedArrow

\u{2b8d}: '⮍'

AnticlockwiseTriangleDashHeadedLeftUDashShapedArrow

\u{2b8e}: '⮎'

AnticlockwiseTriangleDashHeadedTopUDashShapedArrow

\u{2b8f}: '⮏'

ReturnLeft

\u{2b90}: '⮐'

ReturnRight

\u{2b91}: '⮑'

NewlineLeft

\u{2b92}: '⮒'

NewlineRight

\u{2b93}: '⮓'

FourCornerArrowsCirclingAnticlockwise

\u{2b94}: '⮔'

RightwardsBlackArrow

\u{2b95}: '⮕'

ThreeDashDTopDashLightedLeftwardsEquilateralArrowhead

\u{2b98}: '⮘'

ThreeDashDRightDashLightedUpwardsEquilateralArrowhead

\u{2b99}: '⮙'

ThreeDashDTopDashLightedRightwardsEquilateralArrowhead

\u{2b9a}: '⮚'

ThreeDashDLeftDashLightedDownwardsEquilateralArrowhead

\u{2b9b}: '⮛'

BlackLeftwardsEquilateralArrowhead

\u{2b9c}: '⮜'

BlackUpwardsEquilateralArrowhead

\u{2b9d}: '⮝'

BlackRightwardsEquilateralArrowhead

\u{2b9e}: '⮞'

BlackDownwardsEquilateralArrowhead

\u{2b9f}: '⮟'

DownwardsTriangleDashHeadedArrowWithLongTipLeftwards

\u{2ba0}: '⮠'

DownwardsTriangleDashHeadedArrowWithLongTipRightwards

\u{2ba1}: '⮡'

UpwardsTriangleDashHeadedArrowWithLongTipLeftwards

\u{2ba2}: '⮢'

UpwardsTriangleDashHeadedArrowWithLongTipRightwards

\u{2ba3}: '⮣'

LeftwardsTriangleDashHeadedArrowWithLongTipUpwards

\u{2ba4}: '⮤'

RightwardsTriangleDashHeadedArrowWithLongTipUpwards

\u{2ba5}: '⮥'

LeftwardsTriangleDashHeadedArrowWithLongTipDownwards

\u{2ba6}: '⮦'

RightwardsTriangleDashHeadedArrowWithLongTipDownwards

\u{2ba7}: '⮧'

BlackCurvedDownwardsAndLeftwardsArrow

\u{2ba8}: '⮨'

BlackCurvedDownwardsAndRightwardsArrow

\u{2ba9}: '⮩'

BlackCurvedUpwardsAndLeftwardsArrow

\u{2baa}: '⮪'

BlackCurvedUpwardsAndRightwardsArrow

\u{2bab}: '⮫'

BlackCurvedLeftwardsAndUpwardsArrow

\u{2bac}: '⮬'

BlackCurvedRightwardsAndUpwardsArrow

\u{2bad}: '⮭'

BlackCurvedLeftwardsAndDownwardsArrow

\u{2bae}: '⮮'

BlackCurvedRightwardsAndDownwardsArrow

\u{2baf}: '⮯'

RibbonArrowDownLeft

\u{2bb0}: '⮰'

RibbonArrowDownRight

\u{2bb1}: '⮱'

RibbonArrowUpLeft

\u{2bb2}: '⮲'

RibbonArrowUpRight

\u{2bb3}: '⮳'

RibbonArrowLeftUp

\u{2bb4}: '⮴'

RibbonArrowRightUp

\u{2bb5}: '⮵'

RibbonArrowLeftDown

\u{2bb6}: '⮶'

RibbonArrowRightDown

\u{2bb7}: '⮷'

UpwardsWhiteArrowFromBarWithHorizontalBar

\u{2bb8}: '⮸'

UpArrowheadInARectangleBox

\u{2bb9}: '⮹'

OverlappingWhiteSquares

\u{2bba}: '⮺'

OverlappingWhiteAndBlackSquares

\u{2bbb}: '⮻'

OverlappingBlackSquares

\u{2bbc}: '⮼'

BallotBoxWithLightX

\u{2bbd}: '⮽'

CircledX

\u{2bbe}: '⮾'

CircledBoldX

\u{2bbf}: '⮿'

BlackSquareCentred

\u{2bc0}: '⯀'

BlackDiamondCentred

\u{2bc1}: '⯁'

TurnedBlackPentagon

\u{2bc2}: '⯂'

HorizontalBlackOctagon

\u{2bc3}: '⯃'

BlackOctagon

\u{2bc4}: '⯄'

BlackMediumUpDashPointingTriangleCentred

\u{2bc5}: '⯅'

BlackMediumDownDashPointingTriangleCentred

\u{2bc6}: '⯆'

BlackMediumLeftDashPointingTriangleCentred

\u{2bc7}: '⯇'

BlackMediumRightDashPointingTriangleCentred

\u{2bc8}: '⯈'

NeptuneFormTwo

\u{2bc9}: '⯉'

TopHalfBlackCircle

\u{2bca}: '⯊'

BottomHalfBlackCircle

\u{2bcb}: '⯋'

LightFourPointedBlackCusp

\u{2bcc}: '⯌'

RotatedLightFourPointedBlackCusp

\u{2bcd}: '⯍'

WhiteFourPointedCusp

\u{2bce}: '⯎'

RotatedWhiteFourPointedCusp

\u{2bcf}: '⯏'

SquarePositionIndicator

\u{2bd0}: '⯐'

UncertaintySign

\u{2bd1}: '⯑'

GroupMark

\u{2bd2}: '⯒'

PlutoFormTwo

\u{2bd3}: '⯓'

PlutoFormThree

\u{2bd4}: '⯔'

PlutoFormFour

\u{2bd5}: '⯕'

PlutoFormFive

\u{2bd6}: '⯖'

Transpluto

\u{2bd7}: '⯗'

Proserpina

\u{2bd8}: '⯘'

Astraea

\u{2bd9}: '⯙'

Hygiea

\u{2bda}: '⯚'

Pholus

\u{2bdb}: '⯛'

Nessus

\u{2bdc}: '⯜'

WhiteMoonSelena

\u{2bdd}: '⯝'

BlackDiamondOnCross

\u{2bde}: '⯞'

TrueLightMoonArta

\u{2bdf}: '⯟'

Cupido

\u{2be0}: '⯠'

Hades

\u{2be1}: '⯡'

Zeus

\u{2be2}: '⯢'

Kronos

\u{2be3}: '⯣'

Apollon

\u{2be4}: '⯤'

Admetos

\u{2be5}: '⯥'

Vulcanus

\u{2be6}: '⯦'

Poseidon

\u{2be7}: '⯧'

LeftHalfBlackStar

\u{2be8}: '⯨'

RightHalfBlackStar

\u{2be9}: '⯩'

StarWithLeftHalfBlack

\u{2bea}: '⯪'

StarWithRightHalfBlack

\u{2beb}: '⯫'

LeftwardsTwoDashHeadedArrowWithTriangleArrowheads

\u{2bec}: '⯬'

UpwardsTwoDashHeadedArrowWithTriangleArrowheads

\u{2bed}: '⯭'

RightwardsTwoDashHeadedArrowWithTriangleArrowheads

\u{2bee}: '⯮'

DownwardsTwoDashHeadedArrowWithTriangleArrowheads

\u{2bef}: '⯯'

ErisFormOne

\u{2bf0}: '⯰'

ErisFormTwo

\u{2bf1}: '⯱'

Sedna

\u{2bf2}: '⯲'

RussianAstrologicalSymbolVigintile

\u{2bf3}: '⯳'

RussianAstrologicalSymbolNovile

\u{2bf4}: '⯴'

RussianAstrologicalSymbolQuintile

\u{2bf5}: '⯵'

RussianAstrologicalSymbolBinovile

\u{2bf6}: '⯶'

RussianAstrologicalSymbolSentagon

\u{2bf7}: '⯷'

RussianAstrologicalSymbolTredecile

\u{2bf8}: '⯸'

EqualsSignWithInfinityBelow

\u{2bf9}: '⯹'

UnitedSymbol

\u{2bfa}: '⯺'

SeparatedSymbol

\u{2bfb}: '⯻'

DoubledSymbol

\u{2bfc}: '⯼'

PassedSymbol

\u{2bfd}: '⯽'

ReversedRightAngle

\u{2bfe}: '⯾'

Methods

impl MiscellaneousSymbolsandArrows

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case