Enum unicode_types::miscellaneous_symbols_and_pictographs::MiscellaneousSymbolsandPictographs

pub enum MiscellaneousSymbolsandPictographs {
    Cyclone,
    Foggy,
    ClosedUmbrella,
    NightWithStars,
    SunriseOverMountains,
    Sunrise,
    CityscapeAtDusk,
    SunsetOverBuildings,
    Rainbow,
    BridgeAtNight,
    WaterWave,
    Volcano,
    MilkyWay,
    EarthGlobeEuropeDashAfrica,
    EarthGlobeAmericas,
    EarthGlobeAsiaDashAustralia,
    GlobeWithMeridians,
    NewMoonSymbol,
    WaxingCrescentMoonSymbol,
    FirstQuarterMoonSymbol,
    WaxingGibbousMoonSymbol,
    FullMoonSymbol,
    WaningGibbousMoonSymbol,
    LastQuarterMoonSymbol,
    WaningCrescentMoonSymbol,
    CrescentMoon,
    NewMoonWithFace,
    FirstQuarterMoonWithFace,
    LastQuarterMoonWithFace,
    FullMoonWithFace,
    SunWithFace,
    GlowingStar,
    ShootingStar,
    Thermometer,
    BlackDroplet,
    WhiteSun,
    WhiteSunWithSmallCloud,
    WhiteSunBehindCloud,
    WhiteSunBehindCloudWithRain,
    CloudWithRain,
    CloudWithSnow,
    CloudWithLightning,
    CloudWithTornado,
    Fog,
    WindBlowingFace,
    HotDog,
    Taco,
    Burrito,
    Chestnut,
    Seedling,
    EvergreenTree,
    DeciduousTree,
    PalmTree,
    Cactus,
    HotPepper,
    Tulip,
    CherryBlossom,
    Rose,
    Hibiscus,
    Sunflower,
    Blossom,
    EarOfMaize,
    EarOfRice,
    Herb,
    FourLeafClover,
    MapleLeaf,
    FallenLeaf,
    LeafFlutteringInWind,
    Mushroom,
    Tomato,
    Aubergine,
    Grapes,
    Melon,
    Watermelon,
    Tangerine,
    Lemon,
    Banana,
    Pineapple,
    RedApple,
    GreenApple,
    Pear,
    Peach,
    Cherries,
    Strawberry,
    Hamburger,
    SliceOfPizza,
    MeatOnBone,
    PoultryLeg,
    RiceCracker,
    RiceBall,
    CookedRice,
    CurryAndRice,
    SteamingBowl,
    Spaghetti,
    Bread,
    FrenchFries,
    RoastedSweetPotato,
    Dango,
    Oden,
    Sushi,
    FriedShrimp,
    FishCakeWithSwirlDesign,
    SoftIceCream,
    ShavedIce,
    IceCream,
    Doughnut,
    Cookie,
    ChocolateBar,
    Candy,
    Lollipop,
    Custard,
    HoneyPot,
    Shortcake,
    BentoBox,
    PotOfFood,
    Cooking,
    ForkAndKnife,
    TeacupWithoutHandle,
    SakeBottleAndCup,
    WineGlass,
    CocktailGlass,
    TropicalDrink,
    BeerMug,
    ClinkingBeerMugs,
    BabyBottle,
    ForkAndKnifeWithPlate,
    BottleWithPoppingCork,
    Popcorn,
    Ribbon,
    WrappedPresent,
    BirthdayCake,
    JackDashODashLantern,
    ChristmasTree,
    FatherChristmas,
    Fireworks,
    FireworkSparkler,
    Balloon,
    PartyPopper,
    ConfettiBall,
    TanabataTree,
    CrossedFlags,
    PineDecoration,
    JapaneseDolls,
    CarpStreamer,
    WindChime,
    MoonViewingCeremony,
    SchoolSatchel,
    GraduationCap,
    HeartWithTipOnTheLeft,
    BouquetOfFlowers,
    MilitaryMedal,
    ReminderRibbon,
    MusicalKeyboardWithJacks,
    StudioMicrophone,
    LevelSlider,
    ControlKnobs,
    BeamedAscendingMusicalNotes,
    BeamedDescendingMusicalNotes,
    FilmFrames,
    AdmissionTickets,
    CarouselHorse,
    FerrisWheel,
    RollerCoaster,
    FishingPoleAndFish,
    Microphone,
    MovieCamera,
    Cinema,
    Headphone,
    ArtistPalette,
    TopHat,
    CircusTent,
    Ticket,
    ClapperBoard,
    PerformingArts,
    VideoGame,
    DirectHit,
    SlotMachine,
    Billiards,
    GameDie,
    Bowling,
    FlowerPlayingCards,
    MusicalNote,
    MultipleMusicalNotes,
    Saxophone,
    Guitar,
    MusicalKeyboard,
    Trumpet,
    Violin,
    MusicalScore,
    RunningShirtWithSash,
    TennisRacquetAndBall,
    SkiAndSkiBoot,
    BasketballAndHoop,
    ChequeredFlag,
    Snowboarder,
    Runner,
    Surfer,
    SportsMedal,
    Trophy,
    HorseRacing,
    AmericanFootball,
    RugbyFootball,
    Swimmer,
    WeightLifter,
    Golfer,
    RacingMotorcycle,
    RacingCar,
    CricketBatAndBall,
    Volleyball,
    FieldHockeyStickAndBall,
    IceHockeyStickAndPuck,
    TableTennisPaddleAndBall,
    SnowCappedMountain,
    Camping,
    BeachWithUmbrella,
    BuildingConstruction,
    HouseBuildings,
    Cityscape,
    DerelictHouseBuilding,
    ClassicalBuilding,
    Desert,
    DesertIsland,
    NationalPark,
    Stadium,
    HouseBuilding,
    HouseWithGarden,
    OfficeBuilding,
    JapanesePostOffice,
    EuropeanPostOffice,
    Hospital,
    Bank,
    AutomatedTellerMachine,
    Hotel,
    LoveHotel,
    ConvenienceStore,
    School,
    DepartmentStore,
    Factory,
    IzakayaLantern,
    JapaneseCastle,
    EuropeanCastle,
    WhitePennant,
    BlackPennant,
    WavingWhiteFlag,
    WavingBlackFlag,
    Rosette,
    BlackRosette,
    Label,
    BadmintonRacquetAndShuttlecock,
    BowAndArrow,
    Amphora,
    EmojiModifierFitzpatrickTypeDash1Dash2,
    EmojiModifierFitzpatrickTypeDash3,
    EmojiModifierFitzpatrickTypeDash4,
    EmojiModifierFitzpatrickTypeDash5,
    EmojiModifierFitzpatrickTypeDash6,
    Rat,
    Mouse,
    Ox,
    WaterBuffalo,
    Cow,
    Tiger,
    Leopard,
    Rabbit,
    Cat,
    Dragon,
    Crocodile,
    Whale,
    Snail,
    Snake,
    Horse,
    Ram,
    Goat,
    Sheep,
    Monkey,
    Rooster,
    Chicken,
    Dog,
    Pig,
    Boar,
    Elephant,
    Octopus,
    SpiralShell,
    Bug,
    Ant,
    Honeybee,
    LadyBeetle,
    Fish,
    TropicalFish,
    Blowfish,
    Turtle,
    HatchingChick,
    BabyChick,
    FrontDashFacingBabyChick,
    Bird,
    Penguin,
    Koala,
    Poodle,
    DromedaryCamel,
    BactrianCamel,
    Dolphin,
    MouseFace,
    CowFace,
    TigerFace,
    RabbitFace,
    CatFace,
    DragonFace,
    SpoutingWhale,
    HorseFace,
    MonkeyFace,
    DogFace,
    PigFace,
    FrogFace,
    HamsterFace,
    WolfFace,
    BearFace,
    PandaFace,
    PigNose,
    PawPrints,
    Chipmunk,
    Eyes,
    Eye,
    Ear,
    Nose,
    Mouth,
    Tongue,
    WhiteUpPointingBackhandIndex,
    WhiteDownPointingBackhandIndex,
    WhiteLeftPointingBackhandIndex,
    WhiteRightPointingBackhandIndex,
    FistedHandSign,
    WavingHandSign,
    OkHandSign,
    ThumbsUpSign,
    ThumbsDownSign,
    ClappingHandsSign,
    OpenHandsSign,
    Crown,
    WomansHat,
    Eyeglasses,
    Necktie,
    TDashShirt,
    Jeans,
    Dress,
    Kimono,
    Bikini,
    WomansClothes,
    Purse,
    Handbag,
    Pouch,
    MansShoe,
    AthleticShoe,
    HighDashHeeledShoe,
    WomansSandal,
    WomansBoots,
    Footprints,
    BustInSilhouette,
    BustsInSilhouette,
    Boy,
    Girl,
    Man,
    Woman,
    Family,
    ManAndWomanHoldingHands,
    TwoMenHoldingHands,
    TwoWomenHoldingHands,
    PoliceOfficer,
    WomanWithBunnyEars,
    BrideWithVeil,
    PersonWithBlondHair,
    ManWithGuaPiMao,
    ManWithTurban,
    OlderMan,
    OlderWoman,
    Baby,
    ConstructionWorker,
    Princess,
    JapaneseOgre,
    JapaneseGoblin,
    Ghost,
    BabyAngel,
    ExtraterrestrialAlien,
    AlienMonster,
    Imp,
    Skull,
    InformationDeskPerson,
    Guardsman,
    Dancer,
    Lipstick,
    NailPolish,
    FaceMassage,
    Haircut,
    BarberPole,
    Syringe,
    Pill,
    KissMark,
    LoveLetter,
    Ring,
    GemStone,
    Kiss,
    Bouquet,
    CoupleWithHeart,
    Wedding,
    BeatingHeart,
    BrokenHeart,
    TwoHearts,
    SparklingHeart,
    GrowingHeart,
    HeartWithArrow,
    BlueHeart,
    GreenHeart,
    YellowHeart,
    PurpleHeart,
    HeartWithRibbon,
    RevolvingHearts,
    HeartDecoration,
    DiamondShapeWithADotInside,
    ElectricLightBulb,
    AngerSymbol,
    Bomb,
    SleepingSymbol,
    CollisionSymbol,
    SplashingSweatSymbol,
    Droplet,
    DashSymbol,
    PileOfPoo,
    FlexedBiceps,
    DizzySymbol,
    SpeechBalloon,
    ThoughtBalloon,
    WhiteFlower,
    HundredPointsSymbol,
    MoneyBag,
    CurrencyExchange,
    HeavyDollarSign,
    CreditCard,
    BanknoteWithYenSign,
    BanknoteWithDollarSign,
    BanknoteWithEuroSign,
    BanknoteWithPoundSign,
    MoneyWithWings,
    ChartWithUpwardsTrendAndYenSign,
    Seat,
    PersonalComputer,
    Briefcase,
    Minidisc,
    FloppyDisk,
    OpticalDisc,
    Dvd,
    FileFolder,
    OpenFileFolder,
    PageWithCurl,
    PageFacingUp,
    Calendar,
    TearDashOffCalendar,
    CardIndex,
    ChartWithUpwardsTrend,
    ChartWithDownwardsTrend,
    BarChart,
    Clipboard,
    Pushpin,
    RoundPushpin,
    Paperclip,
    StraightRuler,
    TriangularRuler,
    BookmarkTabs,
    Ledger,
    Notebook,
    NotebookWithDecorativeCover,
    ClosedBook,
    OpenBook,
    GreenBook,
    BlueBook,
    OrangeBook,
    Books,
    NameBadge,
    Scroll,
    Memo,
    TelephoneReceiver,
    Pager,
    FaxMachine,
    SatelliteAntenna,
    PublicAddressLoudspeaker,
    CheeringMegaphone,
    OutboxTray,
    InboxTray,
    Package,
    EDashMailSymbol,
    IncomingEnvelope,
    EnvelopeWithDownwardsArrowAbove,
    ClosedMailboxWithLoweredFlag,
    ClosedMailboxWithRaisedFlag,
    OpenMailboxWithRaisedFlag,
    OpenMailboxWithLoweredFlag,
    Postbox,
    PostalHorn,
    Newspaper,
    MobilePhone,
    MobilePhoneWithRightwardsArrowAtLeft,
    VibrationMode,
    MobilePhoneOff,
    NoMobilePhones,
    AntennaWithBars,
    Camera,
    CameraWithFlash,
    VideoCamera,
    Television,
    Radio,
    Videocassette,
    FilmProjector,
    PortableStereo,
    PrayerBeads,
    TwistedRightwardsArrows,
    ClockwiseRightwardsAndLeftwardsOpenCircleArrows,
    ClockwiseRightwardsAndLeftwardsOpenCircleArrowsWithCircledOneOverlay,
    ClockwiseDownwardsAndUpwardsOpenCircleArrows,
    AnticlockwiseDownwardsAndUpwardsOpenCircleArrows,
    LowBrightnessSymbol,
    HighBrightnessSymbol,
    SpeakerWithCancellationStroke,
    Speaker,
    SpeakerWithOneSoundWave,
    SpeakerWithThreeSoundWaves,
    Battery,
    ElectricPlug,
    LeftDashPointingMagnifyingGlass,
    RightDashPointingMagnifyingGlass,
    LockWithInkPen,
    ClosedLockWithKey,
    Key,
    Lock,
    OpenLock,
    Bell,
    BellWithCancellationStroke,
    Bookmark,
    LinkSymbol,
    RadioButton,
    BackWithLeftwardsArrowAbove,
    EndWithLeftwardsArrowAbove,
    OnWithExclamationMarkWithLeftRightArrowAbove,
    SoonWithRightwardsArrowAbove,
    TopWithUpwardsArrowAbove,
    NoOneUnderEighteenSymbol,
    KeycapTen,
    InputSymbolForLatinCapitalLetters,
    InputSymbolForLatinSmallLetters,
    InputSymbolForNumbers,
    InputSymbolForSymbols,
    InputSymbolForLatinLetters,
    Fire,
    ElectricTorch,
    Wrench,
    Hammer,
    NutAndBolt,
    Hocho,
    Pistol,
    Microscope,
    Telescope,
    CrystalBall,
    SixPointedStarWithMiddleDot,
    JapaneseSymbolForBeginner,
    TridentEmblem,
    BlackSquareButton,
    WhiteSquareButton,
    LargeRedCircle,
    LargeBlueCircle,
    LargeOrangeDiamond,
    LargeBlueDiamond,
    SmallOrangeDiamond,
    SmallBlueDiamond,
    UpDashPointingRedTriangle,
    DownDashPointingRedTriangle,
    UpDashPointingSmallRedTriangle,
    DownDashPointingSmallRedTriangle,
    LowerRightShadowedWhiteCircle,
    UpperRightShadowedWhiteCircle,
    CircledCrossPommee,
    CrossPommeeWithHalfDashCircleBelow,
    CrossPommee,
    NotchedLeftSemicircleWithThreeDots,
    NotchedRightSemicircleWithThreeDots,
    SymbolForMarksChapter,
    WhiteLatinCross,
    HeavyLatinCross,
    CelticCross,
    OmSymbol,
    DoveOfPeace,
    Kaaba,
    Mosque,
    Synagogue,
    MenorahWithNineBranches,
    BowlOfHygieia,
    ClockFaceOneOclock,
    ClockFaceTwoOclock,
    ClockFaceThreeOclock,
    ClockFaceFourOclock,
    ClockFaceFiveOclock,
    ClockFaceSixOclock,
    ClockFaceSevenOclock,
    ClockFaceEightOclock,
    ClockFaceNineOclock,
    ClockFaceTenOclock,
    ClockFaceElevenOclock,
    ClockFaceTwelveOclock,
    ClockFaceOneDashThirty,
    ClockFaceTwoDashThirty,
    ClockFaceThreeDashThirty,
    ClockFaceFourDashThirty,
    ClockFaceFiveDashThirty,
    ClockFaceSixDashThirty,
    ClockFaceSevenDashThirty,
    ClockFaceEightDashThirty,
    ClockFaceNineDashThirty,
    ClockFaceTenDashThirty,
    ClockFaceElevenDashThirty,
    ClockFaceTwelveDashThirty,
    RightSpeaker,
    RightSpeakerWithOneSoundWave,
    RightSpeakerWithThreeSoundWaves,
    Bullhorn,
    BullhornWithSoundWaves,
    RingingBell,
    Book,
    Candle,
    MantelpieceClock,
    BlackSkullAndCrossbones,
    NoPiracy,
    Hole,
    ManInBusinessSuitLevitating,
    SleuthOrSpy,
    DarkSunglasses,
    Spider,
    SpiderWeb,
    Joystick,
    ManDancing,
    LeftHandTelephoneReceiver,
    TelephoneReceiverWithPage,
    RightHandTelephoneReceiver,
    WhiteTouchtoneTelephone,
    BlackTouchtoneTelephone,
    TelephoneOnTopOfModem,
    ClamshellMobilePhone,
    BackOfEnvelope,
    StampedEnvelope,
    EnvelopeWithLightning,
    FlyingEnvelope,
    PenOverStampedEnvelope,
    LinkedPaperclips,
    BlackPushpin,
    LowerLeftPencil,
    LowerLeftBallpointPen,
    LowerLeftFountainPen,
    LowerLeftPaintbrush,
    LowerLeftCrayon,
    LeftWritingHand,
    TurnedOkHandSign,
    RaisedHandWithFingersSplayed,
    ReversedRaisedHandWithFingersSplayed,
    ReversedThumbsUpSign,
    ReversedThumbsDownSign,
    ReversedVictoryHand,
    ReversedHandWithMiddleFingerExtended,
    RaisedHandWithPartBetweenMiddleAndRingFingers,
    WhiteDownPointingLeftHandIndex,
    SidewaysWhiteLeftPointingIndex,
    SidewaysWhiteRightPointingIndex,
    SidewaysBlackLeftPointingIndex,
    SidewaysBlackRightPointingIndex,
    BlackLeftPointingBackhandIndex,
    BlackRightPointingBackhandIndex,
    SidewaysWhiteUpPointingIndex,
    SidewaysWhiteDownPointingIndex,
    SidewaysBlackUpPointingIndex,
    SidewaysBlackDownPointingIndex,
    BlackUpPointingBackhandIndex,
    BlackDownPointingBackhandIndex,
    BlackHeart,
    DesktopComputer,
    KeyboardAndMouse,
    ThreeNetworkedComputers,
    Printer,
    PocketCalculator,
    BlackHardShellFloppyDisk,
    WhiteHardShellFloppyDisk,
    SoftShellFloppyDisk,
    TapeCartridge,
    WiredKeyboard,
    OneButtonMouse,
    TwoButtonMouse,
    ThreeButtonMouse,
    Trackball,
    OldPersonalComputer,
    HardDisk,
    Screen,
    PrinterIcon,
    FaxIcon,
    OpticalDiscIcon,
    DocumentWithText,
    DocumentWithTextAndPicture,
    DocumentWithPicture,
    FrameWithPicture,
    FrameWithTiles,
    FrameWithAnX,
    BlackFolder,
    Folder,
    OpenFolder,
    CardIndexDividers,
    CardFileBox,
    FileCabinet,
    EmptyNote,
    EmptyNotePage,
    EmptyNotePad,
    Note,
    NotePage,
    NotePad,
    EmptyDocument,
    EmptyPage,
    EmptyPages,
    Document,
    Page,
    Pages,
    Wastebasket,
    SpiralNotePad,
    SpiralCalendarPad,
    DesktopWindow,
    Minimize,
    Maximize,
    Overlap,
    ClockwiseRightAndLeftSemicircleArrows,
    CancellationX,
    IncreaseFontSizeSymbol,
    DecreaseFontSizeSymbol,
    Compression,
    OldKey,
    RolledDashUpNewspaper,
    PageWithCircledText,
    StockChart,
    DaggerKnife,
    Lips,
    SpeakingHeadInSilhouette,
    ThreeRaysAbove,
    ThreeRaysBelow,
    ThreeRaysLeft,
    ThreeRaysRight,
    LeftSpeechBubble,
    RightSpeechBubble,
    TwoSpeechBubbles,
    ThreeSpeechBubbles,
    LeftThoughtBubble,
    RightThoughtBubble,
    LeftAngerBubble,
    RightAngerBubble,
    MoodBubble,
    LightningMoodBubble,
    LightningMood,
    BallotBoxWithBallot,
    BallotScriptX,
    BallotBoxWithScriptX,
    BallotBoldScriptX,
    BallotBoxWithBoldScriptX,
    LightCheckMark,
    BallotBoxWithBoldCheck,
    WorldMap,
    MountFuji,
    TokyoTower,
    StatueOfLiberty,
    SilhouetteOfJapan,
}

An enum to represent all characters in the MiscellaneousSymbolsandPictographs block.

Variants
Cyclone

\u{1f300}: '🌀'

Foggy

\u{1f301}: '🌁'

ClosedUmbrella

\u{1f302}: '🌂'

NightWithStars

\u{1f303}: '🌃'

SunriseOverMountains

\u{1f304}: '🌄'

Sunrise

\u{1f305}: '🌅'

CityscapeAtDusk

\u{1f306}: '🌆'

SunsetOverBuildings

\u{1f307}: '🌇'

Rainbow

\u{1f308}: '🌈'

BridgeAtNight

\u{1f309}: '🌉'

WaterWave

\u{1f30a}: '🌊'

Volcano

\u{1f30b}: '🌋'

MilkyWay

\u{1f30c}: '🌌'

EarthGlobeEuropeDashAfrica

\u{1f30d}: '🌍'

EarthGlobeAmericas

\u{1f30e}: '🌎'

EarthGlobeAsiaDashAustralia

\u{1f30f}: '🌏'

GlobeWithMeridians

\u{1f310}: '🌐'

NewMoonSymbol

\u{1f311}: '🌑'

WaxingCrescentMoonSymbol

\u{1f312}: '🌒'

FirstQuarterMoonSymbol

\u{1f313}: '🌓'

WaxingGibbousMoonSymbol

\u{1f314}: '🌔'

FullMoonSymbol

\u{1f315}: '🌕'

WaningGibbousMoonSymbol

\u{1f316}: '🌖'

LastQuarterMoonSymbol

\u{1f317}: '🌗'

WaningCrescentMoonSymbol

\u{1f318}: '🌘'

CrescentMoon

\u{1f319}: '🌙'

NewMoonWithFace

\u{1f31a}: '🌚'

FirstQuarterMoonWithFace

\u{1f31b}: '🌛'

LastQuarterMoonWithFace

\u{1f31c}: '🌜'

FullMoonWithFace

\u{1f31d}: '🌝'

SunWithFace

\u{1f31e}: '🌞'

GlowingStar

\u{1f31f}: '🌟'

ShootingStar

\u{1f320}: '🌠'

Thermometer

\u{1f321}: '🌡'

BlackDroplet

\u{1f322}: '🌢'

WhiteSun

\u{1f323}: '🌣'

WhiteSunWithSmallCloud

\u{1f324}: '🌤'

WhiteSunBehindCloud

\u{1f325}: '🌥'

WhiteSunBehindCloudWithRain

\u{1f326}: '🌦'

CloudWithRain

\u{1f327}: '🌧'

CloudWithSnow

\u{1f328}: '🌨'

CloudWithLightning

\u{1f329}: '🌩'

CloudWithTornado

\u{1f32a}: '🌪'

Fog

\u{1f32b}: '🌫'

WindBlowingFace

\u{1f32c}: '🌬'

HotDog

\u{1f32d}: '🌭'

Taco

\u{1f32e}: '🌮'

Burrito

\u{1f32f}: '🌯'

Chestnut

\u{1f330}: '🌰'

Seedling

\u{1f331}: '🌱'

EvergreenTree

\u{1f332}: '🌲'

DeciduousTree

\u{1f333}: '🌳'

PalmTree

\u{1f334}: '🌴'

Cactus

\u{1f335}: '🌵'

HotPepper

\u{1f336}: '🌶'

Tulip

\u{1f337}: '🌷'

CherryBlossom

\u{1f338}: '🌸'

Rose

\u{1f339}: '🌹'

Hibiscus

\u{1f33a}: '🌺'

Sunflower

\u{1f33b}: '🌻'

Blossom

\u{1f33c}: '🌼'

EarOfMaize

\u{1f33d}: '🌽'

EarOfRice

\u{1f33e}: '🌾'

Herb

\u{1f33f}: '🌿'

FourLeafClover

\u{1f340}: '🍀'

MapleLeaf

\u{1f341}: '🍁'

FallenLeaf

\u{1f342}: '🍂'

LeafFlutteringInWind

\u{1f343}: '🍃'

Mushroom

\u{1f344}: '🍄'

Tomato

\u{1f345}: '🍅'

Aubergine

\u{1f346}: '🍆'

Grapes

\u{1f347}: '🍇'

Melon

\u{1f348}: '🍈'

Watermelon

\u{1f349}: '🍉'

Tangerine

\u{1f34a}: '🍊'

Lemon

\u{1f34b}: '🍋'

Banana

\u{1f34c}: '🍌'

Pineapple

\u{1f34d}: '🍍'

RedApple

\u{1f34e}: '🍎'

GreenApple

\u{1f34f}: '🍏'

Pear

\u{1f350}: '🍐'

Peach

\u{1f351}: '🍑'

Cherries

\u{1f352}: '🍒'

Strawberry

\u{1f353}: '🍓'

Hamburger

\u{1f354}: '🍔'

SliceOfPizza

\u{1f355}: '🍕'

MeatOnBone

\u{1f356}: '🍖'

PoultryLeg

\u{1f357}: '🍗'

RiceCracker

\u{1f358}: '🍘'

RiceBall

\u{1f359}: '🍙'

CookedRice

\u{1f35a}: '🍚'

CurryAndRice

\u{1f35b}: '🍛'

SteamingBowl

\u{1f35c}: '🍜'

Spaghetti

\u{1f35d}: '🍝'

Bread

\u{1f35e}: '🍞'

FrenchFries

\u{1f35f}: '🍟'

RoastedSweetPotato

\u{1f360}: '🍠'

Dango

\u{1f361}: '🍡'

Oden

\u{1f362}: '🍢'

Sushi

\u{1f363}: '🍣'

FriedShrimp

\u{1f364}: '🍤'

FishCakeWithSwirlDesign

\u{1f365}: '🍥'

SoftIceCream

\u{1f366}: '🍦'

ShavedIce

\u{1f367}: '🍧'

IceCream

\u{1f368}: '🍨'

Doughnut

\u{1f369}: '🍩'

Cookie

\u{1f36a}: '🍪'

ChocolateBar

\u{1f36b}: '🍫'

Candy

\u{1f36c}: '🍬'

Lollipop

\u{1f36d}: '🍭'

Custard

\u{1f36e}: '🍮'

HoneyPot

\u{1f36f}: '🍯'

Shortcake

\u{1f370}: '🍰'

BentoBox

\u{1f371}: '🍱'

PotOfFood

\u{1f372}: '🍲'

Cooking

\u{1f373}: '🍳'

ForkAndKnife

\u{1f374}: '🍴'

TeacupWithoutHandle

\u{1f375}: '🍵'

SakeBottleAndCup

\u{1f376}: '🍶'

WineGlass

\u{1f377}: '🍷'

CocktailGlass

\u{1f378}: '🍸'

TropicalDrink

\u{1f379}: '🍹'

BeerMug

\u{1f37a}: '🍺'

ClinkingBeerMugs

\u{1f37b}: '🍻'

BabyBottle

\u{1f37c}: '🍼'

ForkAndKnifeWithPlate

\u{1f37d}: '🍽'

BottleWithPoppingCork

\u{1f37e}: '🍾'

Popcorn

\u{1f37f}: '🍿'

Ribbon

\u{1f380}: '🎀'

WrappedPresent

\u{1f381}: '🎁'

BirthdayCake

\u{1f382}: '🎂'

JackDashODashLantern

\u{1f383}: '🎃'

ChristmasTree

\u{1f384}: '🎄'

FatherChristmas

\u{1f385}: '🎅'

Fireworks

\u{1f386}: '🎆'

FireworkSparkler

\u{1f387}: '🎇'

Balloon

\u{1f388}: '🎈'

PartyPopper

\u{1f389}: '🎉'

ConfettiBall

\u{1f38a}: '🎊'

TanabataTree

\u{1f38b}: '🎋'

CrossedFlags

\u{1f38c}: '🎌'

PineDecoration

\u{1f38d}: '🎍'

JapaneseDolls

\u{1f38e}: '🎎'

CarpStreamer

\u{1f38f}: '🎏'

WindChime

\u{1f390}: '🎐'

MoonViewingCeremony

\u{1f391}: '🎑'

SchoolSatchel

\u{1f392}: '🎒'

GraduationCap

\u{1f393}: '🎓'

HeartWithTipOnTheLeft

\u{1f394}: '🎔'

BouquetOfFlowers

\u{1f395}: '🎕'

MilitaryMedal

\u{1f396}: '🎖'

ReminderRibbon

\u{1f397}: '🎗'

MusicalKeyboardWithJacks

\u{1f398}: '🎘'

StudioMicrophone

\u{1f399}: '🎙'

LevelSlider

\u{1f39a}: '🎚'

ControlKnobs

\u{1f39b}: '🎛'

BeamedAscendingMusicalNotes

\u{1f39c}: '🎜'

BeamedDescendingMusicalNotes

\u{1f39d}: '🎝'

FilmFrames

\u{1f39e}: '🎞'

AdmissionTickets

\u{1f39f}: '🎟'

CarouselHorse

\u{1f3a0}: '🎠'

FerrisWheel

\u{1f3a1}: '🎡'

RollerCoaster

\u{1f3a2}: '🎢'

FishingPoleAndFish

\u{1f3a3}: '🎣'

Microphone

\u{1f3a4}: '🎤'

MovieCamera

\u{1f3a5}: '🎥'

Cinema

\u{1f3a6}: '🎦'

Headphone

\u{1f3a7}: '🎧'

ArtistPalette

\u{1f3a8}: '🎨'

TopHat

\u{1f3a9}: '🎩'

CircusTent

\u{1f3aa}: '🎪'

Ticket

\u{1f3ab}: '🎫'

ClapperBoard

\u{1f3ac}: '🎬'

PerformingArts

\u{1f3ad}: '🎭'

VideoGame

\u{1f3ae}: '🎮'

DirectHit

\u{1f3af}: '🎯'

SlotMachine

\u{1f3b0}: '🎰'

Billiards

\u{1f3b1}: '🎱'

GameDie

\u{1f3b2}: '🎲'

Bowling

\u{1f3b3}: '🎳'

FlowerPlayingCards

\u{1f3b4}: '🎴'

MusicalNote

\u{1f3b5}: '🎵'

MultipleMusicalNotes

\u{1f3b6}: '🎶'

Saxophone

\u{1f3b7}: '🎷'

Guitar

\u{1f3b8}: '🎸'

MusicalKeyboard

\u{1f3b9}: '🎹'

Trumpet

\u{1f3ba}: '🎺'

Violin

\u{1f3bb}: '🎻'

MusicalScore

\u{1f3bc}: '🎼'

RunningShirtWithSash

\u{1f3bd}: '🎽'

TennisRacquetAndBall

\u{1f3be}: '🎾'

SkiAndSkiBoot

\u{1f3bf}: '🎿'

BasketballAndHoop

\u{1f3c0}: '🏀'

ChequeredFlag

\u{1f3c1}: '🏁'

Snowboarder

\u{1f3c2}: '🏂'

Runner

\u{1f3c3}: '🏃'

Surfer

\u{1f3c4}: '🏄'

SportsMedal

\u{1f3c5}: '🏅'

Trophy

\u{1f3c6}: '🏆'

HorseRacing

\u{1f3c7}: '🏇'

AmericanFootball

\u{1f3c8}: '🏈'

RugbyFootball

\u{1f3c9}: '🏉'

Swimmer

\u{1f3ca}: '🏊'

WeightLifter

\u{1f3cb}: '🏋'

Golfer

\u{1f3cc}: '🏌'

RacingMotorcycle

\u{1f3cd}: '🏍'

RacingCar

\u{1f3ce}: '🏎'

CricketBatAndBall

\u{1f3cf}: '🏏'

Volleyball

\u{1f3d0}: '🏐'

FieldHockeyStickAndBall

\u{1f3d1}: '🏑'

IceHockeyStickAndPuck

\u{1f3d2}: '🏒'

TableTennisPaddleAndBall

\u{1f3d3}: '🏓'

SnowCappedMountain

\u{1f3d4}: '🏔'

Camping

\u{1f3d5}: '🏕'

BeachWithUmbrella

\u{1f3d6}: '🏖'

BuildingConstruction

\u{1f3d7}: '🏗'

HouseBuildings

\u{1f3d8}: '🏘'

Cityscape

\u{1f3d9}: '🏙'

DerelictHouseBuilding

\u{1f3da}: '🏚'

ClassicalBuilding

\u{1f3db}: '🏛'

Desert

\u{1f3dc}: '🏜'

DesertIsland

\u{1f3dd}: '🏝'

NationalPark

\u{1f3de}: '🏞'

Stadium

\u{1f3df}: '🏟'

HouseBuilding

\u{1f3e0}: '🏠'

HouseWithGarden

\u{1f3e1}: '🏡'

OfficeBuilding

\u{1f3e2}: '🏢'

JapanesePostOffice

\u{1f3e3}: '🏣'

EuropeanPostOffice

\u{1f3e4}: '🏤'

Hospital

\u{1f3e5}: '🏥'

Bank

\u{1f3e6}: '🏦'

AutomatedTellerMachine

\u{1f3e7}: '🏧'

Hotel

\u{1f3e8}: '🏨'

LoveHotel

\u{1f3e9}: '🏩'

ConvenienceStore

\u{1f3ea}: '🏪'

School

\u{1f3eb}: '🏫'

DepartmentStore

\u{1f3ec}: '🏬'

Factory

\u{1f3ed}: '🏭'

IzakayaLantern

\u{1f3ee}: '🏮'

JapaneseCastle

\u{1f3ef}: '🏯'

EuropeanCastle

\u{1f3f0}: '🏰'

WhitePennant

\u{1f3f1}: '🏱'

BlackPennant

\u{1f3f2}: '🏲'

WavingWhiteFlag

\u{1f3f3}: '🏳'

WavingBlackFlag

\u{1f3f4}: '🏴'

Rosette

\u{1f3f5}: '🏵'

BlackRosette

\u{1f3f6}: '🏶'

Label

\u{1f3f7}: '🏷'

BadmintonRacquetAndShuttlecock

\u{1f3f8}: '🏸'

BowAndArrow

\u{1f3f9}: '🏹'

Amphora

\u{1f3fa}: '🏺'

EmojiModifierFitzpatrickTypeDash1Dash2

\u{1f3fb}: '🏻'

EmojiModifierFitzpatrickTypeDash3

\u{1f3fc}: '🏼'

EmojiModifierFitzpatrickTypeDash4

\u{1f3fd}: '🏽'

EmojiModifierFitzpatrickTypeDash5

\u{1f3fe}: '🏾'

EmojiModifierFitzpatrickTypeDash6

\u{1f3ff}: '🏿'

Rat

\u{1f400}: '🐀'

Mouse

\u{1f401}: '🐁'

Ox

\u{1f402}: '🐂'

WaterBuffalo

\u{1f403}: '🐃'

Cow

\u{1f404}: '🐄'

Tiger

\u{1f405}: '🐅'

Leopard

\u{1f406}: '🐆'

Rabbit

\u{1f407}: '🐇'

Cat

\u{1f408}: '🐈'

Dragon

\u{1f409}: '🐉'

Crocodile

\u{1f40a}: '🐊'

Whale

\u{1f40b}: '🐋'

Snail

\u{1f40c}: '🐌'

Snake

\u{1f40d}: '🐍'

Horse

\u{1f40e}: '🐎'

Ram

\u{1f40f}: '🐏'

Goat

\u{1f410}: '🐐'

Sheep

\u{1f411}: '🐑'

Monkey

\u{1f412}: '🐒'

Rooster

\u{1f413}: '🐓'

Chicken

\u{1f414}: '🐔'

Dog

\u{1f415}: '🐕'

Pig

\u{1f416}: '🐖'

Boar

\u{1f417}: '🐗'

Elephant

\u{1f418}: '🐘'

Octopus

\u{1f419}: '🐙'

SpiralShell

\u{1f41a}: '🐚'

Bug

\u{1f41b}: '🐛'

Ant

\u{1f41c}: '🐜'

Honeybee

\u{1f41d}: '🐝'

LadyBeetle

\u{1f41e}: '🐞'

Fish

\u{1f41f}: '🐟'

TropicalFish

\u{1f420}: '🐠'

Blowfish

\u{1f421}: '🐡'

Turtle

\u{1f422}: '🐢'

HatchingChick

\u{1f423}: '🐣'

BabyChick

\u{1f424}: '🐤'

FrontDashFacingBabyChick

\u{1f425}: '🐥'

Bird

\u{1f426}: '🐦'

Penguin

\u{1f427}: '🐧'

Koala

\u{1f428}: '🐨'

Poodle

\u{1f429}: '🐩'

DromedaryCamel

\u{1f42a}: '🐪'

BactrianCamel

\u{1f42b}: '🐫'

Dolphin

\u{1f42c}: '🐬'

MouseFace

\u{1f42d}: '🐭'

CowFace

\u{1f42e}: '🐮'

TigerFace

\u{1f42f}: '🐯'

RabbitFace

\u{1f430}: '🐰'

CatFace

\u{1f431}: '🐱'

DragonFace

\u{1f432}: '🐲'

SpoutingWhale

\u{1f433}: '🐳'

HorseFace

\u{1f434}: '🐴'

MonkeyFace

\u{1f435}: '🐵'

DogFace

\u{1f436}: '🐶'

PigFace

\u{1f437}: '🐷'

FrogFace

\u{1f438}: '🐸'

HamsterFace

\u{1f439}: '🐹'

WolfFace

\u{1f43a}: '🐺'

BearFace

\u{1f43b}: '🐻'

PandaFace

\u{1f43c}: '🐼'

PigNose

\u{1f43d}: '🐽'

PawPrints

\u{1f43e}: '🐾'

Chipmunk

\u{1f43f}: '🐿'

Eyes

\u{1f440}: '👀'

Eye

\u{1f441}: '👁'

Ear

\u{1f442}: '👂'

Nose

\u{1f443}: '👃'

Mouth

\u{1f444}: '👄'

Tongue

\u{1f445}: '👅'

WhiteUpPointingBackhandIndex

\u{1f446}: '👆'

WhiteDownPointingBackhandIndex

\u{1f447}: '👇'

WhiteLeftPointingBackhandIndex

\u{1f448}: '👈'

WhiteRightPointingBackhandIndex

\u{1f449}: '👉'

FistedHandSign

\u{1f44a}: '👊'

WavingHandSign

\u{1f44b}: '👋'

OkHandSign

\u{1f44c}: '👌'

ThumbsUpSign

\u{1f44d}: '👍'

ThumbsDownSign

\u{1f44e}: '👎'

ClappingHandsSign

\u{1f44f}: '👏'

OpenHandsSign

\u{1f450}: '👐'

Crown

\u{1f451}: '👑'

WomansHat

\u{1f452}: '👒'

Eyeglasses

\u{1f453}: '👓'

Necktie

\u{1f454}: '👔'

TDashShirt

\u{1f455}: '👕'

Jeans

\u{1f456}: '👖'

Dress

\u{1f457}: '👗'

Kimono

\u{1f458}: '👘'

Bikini

\u{1f459}: '👙'

WomansClothes

\u{1f45a}: '👚'

Purse

\u{1f45b}: '👛'

Handbag

\u{1f45c}: '👜'

Pouch

\u{1f45d}: '👝'

MansShoe

\u{1f45e}: '👞'

AthleticShoe

\u{1f45f}: '👟'

HighDashHeeledShoe

\u{1f460}: '👠'

WomansSandal

\u{1f461}: '👡'

WomansBoots

\u{1f462}: '👢'

Footprints

\u{1f463}: '👣'

BustInSilhouette

\u{1f464}: '👤'

BustsInSilhouette

\u{1f465}: '👥'

Boy

\u{1f466}: '👦'

Girl

\u{1f467}: '👧'

Man

\u{1f468}: '👨'

Woman

\u{1f469}: '👩'

Family

\u{1f46a}: '👪'

ManAndWomanHoldingHands

\u{1f46b}: '👫'

TwoMenHoldingHands

\u{1f46c}: '👬'

TwoWomenHoldingHands

\u{1f46d}: '👭'

PoliceOfficer

\u{1f46e}: '👮'

WomanWithBunnyEars

\u{1f46f}: '👯'

BrideWithVeil

\u{1f470}: '👰'

PersonWithBlondHair

\u{1f471}: '👱'

ManWithGuaPiMao

\u{1f472}: '👲'

ManWithTurban

\u{1f473}: '👳'

OlderMan

\u{1f474}: '👴'

OlderWoman

\u{1f475}: '👵'

Baby

\u{1f476}: '👶'

ConstructionWorker

\u{1f477}: '👷'

Princess

\u{1f478}: '👸'

JapaneseOgre

\u{1f479}: '👹'

JapaneseGoblin

\u{1f47a}: '👺'

Ghost

\u{1f47b}: '👻'

BabyAngel

\u{1f47c}: '👼'

ExtraterrestrialAlien

\u{1f47d}: '👽'

AlienMonster

\u{1f47e}: '👾'

Imp

\u{1f47f}: '👿'

Skull

\u{1f480}: '💀'

InformationDeskPerson

\u{1f481}: '💁'

Guardsman

\u{1f482}: '💂'

Dancer

\u{1f483}: '💃'

Lipstick

\u{1f484}: '💄'

NailPolish

\u{1f485}: '💅'

FaceMassage

\u{1f486}: '💆'

Haircut

\u{1f487}: '💇'

BarberPole

\u{1f488}: '💈'

Syringe

\u{1f489}: '💉'

Pill

\u{1f48a}: '💊'

KissMark

\u{1f48b}: '💋'

LoveLetter

\u{1f48c}: '💌'

Ring

\u{1f48d}: '💍'

GemStone

\u{1f48e}: '💎'

Kiss

\u{1f48f}: '💏'

Bouquet

\u{1f490}: '💐'

CoupleWithHeart

\u{1f491}: '💑'

Wedding

\u{1f492}: '💒'

BeatingHeart

\u{1f493}: '💓'

BrokenHeart

\u{1f494}: '💔'

TwoHearts

\u{1f495}: '💕'

SparklingHeart

\u{1f496}: '💖'

GrowingHeart

\u{1f497}: '💗'

HeartWithArrow

\u{1f498}: '💘'

BlueHeart

\u{1f499}: '💙'

GreenHeart

\u{1f49a}: '💚'

YellowHeart

\u{1f49b}: '💛'

PurpleHeart

\u{1f49c}: '💜'

HeartWithRibbon

\u{1f49d}: '💝'

RevolvingHearts

\u{1f49e}: '💞'

HeartDecoration

\u{1f49f}: '💟'

DiamondShapeWithADotInside

\u{1f4a0}: '💠'

ElectricLightBulb

\u{1f4a1}: '💡'

AngerSymbol

\u{1f4a2}: '💢'

Bomb

\u{1f4a3}: '💣'

SleepingSymbol

\u{1f4a4}: '💤'

CollisionSymbol

\u{1f4a5}: '💥'

SplashingSweatSymbol

\u{1f4a6}: '💦'

Droplet

\u{1f4a7}: '💧'

DashSymbol

\u{1f4a8}: '💨'

PileOfPoo

\u{1f4a9}: '💩'

FlexedBiceps

\u{1f4aa}: '💪'

DizzySymbol

\u{1f4ab}: '💫'

SpeechBalloon

\u{1f4ac}: '💬'

ThoughtBalloon

\u{1f4ad}: '💭'

WhiteFlower

\u{1f4ae}: '💮'

HundredPointsSymbol

\u{1f4af}: '💯'

MoneyBag

\u{1f4b0}: '💰'

CurrencyExchange

\u{1f4b1}: '💱'

HeavyDollarSign

\u{1f4b2}: '💲'

CreditCard

\u{1f4b3}: '💳'

BanknoteWithYenSign

\u{1f4b4}: '💴'

BanknoteWithDollarSign

\u{1f4b5}: '💵'

BanknoteWithEuroSign

\u{1f4b6}: '💶'

BanknoteWithPoundSign

\u{1f4b7}: '💷'

MoneyWithWings

\u{1f4b8}: '💸'

ChartWithUpwardsTrendAndYenSign

\u{1f4b9}: '💹'

Seat

\u{1f4ba}: '💺'

PersonalComputer

\u{1f4bb}: '💻'

Briefcase

\u{1f4bc}: '💼'

Minidisc

\u{1f4bd}: '💽'

FloppyDisk

\u{1f4be}: '💾'

OpticalDisc

\u{1f4bf}: '💿'

Dvd

\u{1f4c0}: '📀'

FileFolder

\u{1f4c1}: '📁'

OpenFileFolder

\u{1f4c2}: '📂'

PageWithCurl

\u{1f4c3}: '📃'

PageFacingUp

\u{1f4c4}: '📄'

Calendar

\u{1f4c5}: '📅'

TearDashOffCalendar

\u{1f4c6}: '📆'

CardIndex

\u{1f4c7}: '📇'

ChartWithUpwardsTrend

\u{1f4c8}: '📈'

ChartWithDownwardsTrend

\u{1f4c9}: '📉'

BarChart

\u{1f4ca}: '📊'

Clipboard

\u{1f4cb}: '📋'

Pushpin

\u{1f4cc}: '📌'

RoundPushpin

\u{1f4cd}: '📍'

Paperclip

\u{1f4ce}: '📎'

StraightRuler

\u{1f4cf}: '📏'

TriangularRuler

\u{1f4d0}: '📐'

BookmarkTabs

\u{1f4d1}: '📑'

Ledger

\u{1f4d2}: '📒'

Notebook

\u{1f4d3}: '📓'

NotebookWithDecorativeCover

\u{1f4d4}: '📔'

ClosedBook

\u{1f4d5}: '📕'

OpenBook

\u{1f4d6}: '📖'

GreenBook

\u{1f4d7}: '📗'

BlueBook

\u{1f4d8}: '📘'

OrangeBook

\u{1f4d9}: '📙'

Books

\u{1f4da}: '📚'

NameBadge

\u{1f4db}: '📛'

Scroll

\u{1f4dc}: '📜'

Memo

\u{1f4dd}: '📝'

TelephoneReceiver

\u{1f4de}: '📞'

Pager

\u{1f4df}: '📟'

FaxMachine

\u{1f4e0}: '📠'

SatelliteAntenna

\u{1f4e1}: '📡'

PublicAddressLoudspeaker

\u{1f4e2}: '📢'

CheeringMegaphone

\u{1f4e3}: '📣'

OutboxTray

\u{1f4e4}: '📤'

InboxTray

\u{1f4e5}: '📥'

Package

\u{1f4e6}: '📦'

EDashMailSymbol

\u{1f4e7}: '📧'

IncomingEnvelope

\u{1f4e8}: '📨'

EnvelopeWithDownwardsArrowAbove

\u{1f4e9}: '📩'

ClosedMailboxWithLoweredFlag

\u{1f4ea}: '📪'

ClosedMailboxWithRaisedFlag

\u{1f4eb}: '📫'

OpenMailboxWithRaisedFlag

\u{1f4ec}: '📬'

OpenMailboxWithLoweredFlag

\u{1f4ed}: '📭'

Postbox

\u{1f4ee}: '📮'

PostalHorn

\u{1f4ef}: '📯'

Newspaper

\u{1f4f0}: '📰'

MobilePhone

\u{1f4f1}: '📱'

MobilePhoneWithRightwardsArrowAtLeft

\u{1f4f2}: '📲'

VibrationMode

\u{1f4f3}: '📳'

MobilePhoneOff

\u{1f4f4}: '📴'

NoMobilePhones

\u{1f4f5}: '📵'

AntennaWithBars

\u{1f4f6}: '📶'

Camera

\u{1f4f7}: '📷'

CameraWithFlash

\u{1f4f8}: '📸'

VideoCamera

\u{1f4f9}: '📹'

Television

\u{1f4fa}: '📺'

Radio

\u{1f4fb}: '📻'

Videocassette

\u{1f4fc}: '📼'

FilmProjector

\u{1f4fd}: '📽'

PortableStereo

\u{1f4fe}: '📾'

PrayerBeads

\u{1f4ff}: '📿'

TwistedRightwardsArrows

\u{1f500}: '🔀'

ClockwiseRightwardsAndLeftwardsOpenCircleArrows

\u{1f501}: '🔁'

ClockwiseRightwardsAndLeftwardsOpenCircleArrowsWithCircledOneOverlay

\u{1f502}: '🔂'

ClockwiseDownwardsAndUpwardsOpenCircleArrows

\u{1f503}: '🔃'

AnticlockwiseDownwardsAndUpwardsOpenCircleArrows

\u{1f504}: '🔄'

LowBrightnessSymbol

\u{1f505}: '🔅'

HighBrightnessSymbol

\u{1f506}: '🔆'

SpeakerWithCancellationStroke

\u{1f507}: '🔇'

Speaker

\u{1f508}: '🔈'

SpeakerWithOneSoundWave

\u{1f509}: '🔉'

SpeakerWithThreeSoundWaves

\u{1f50a}: '🔊'

Battery

\u{1f50b}: '🔋'

ElectricPlug

\u{1f50c}: '🔌'

LeftDashPointingMagnifyingGlass

\u{1f50d}: '🔍'

RightDashPointingMagnifyingGlass

\u{1f50e}: '🔎'

LockWithInkPen

\u{1f50f}: '🔏'

ClosedLockWithKey

\u{1f510}: '🔐'

Key

\u{1f511}: '🔑'

Lock

\u{1f512}: '🔒'

OpenLock

\u{1f513}: '🔓'

Bell

\u{1f514}: '🔔'

BellWithCancellationStroke

\u{1f515}: '🔕'

Bookmark

\u{1f516}: '🔖'

LinkSymbol

\u{1f517}: '🔗'

RadioButton

\u{1f518}: '🔘'

BackWithLeftwardsArrowAbove

\u{1f519}: '🔙'

EndWithLeftwardsArrowAbove

\u{1f51a}: '🔚'

OnWithExclamationMarkWithLeftRightArrowAbove

\u{1f51b}: '🔛'

SoonWithRightwardsArrowAbove

\u{1f51c}: '🔜'

TopWithUpwardsArrowAbove

\u{1f51d}: '🔝'

NoOneUnderEighteenSymbol

\u{1f51e}: '🔞'

KeycapTen

\u{1f51f}: '🔟'

InputSymbolForLatinCapitalLetters

\u{1f520}: '🔠'

InputSymbolForLatinSmallLetters

\u{1f521}: '🔡'

InputSymbolForNumbers

\u{1f522}: '🔢'

InputSymbolForSymbols

\u{1f523}: '🔣'

InputSymbolForLatinLetters

\u{1f524}: '🔤'

Fire

\u{1f525}: '🔥'

ElectricTorch

\u{1f526}: '🔦'

Wrench

\u{1f527}: '🔧'

Hammer

\u{1f528}: '🔨'

NutAndBolt

\u{1f529}: '🔩'

Hocho

\u{1f52a}: '🔪'

Pistol

\u{1f52b}: '🔫'

Microscope

\u{1f52c}: '🔬'

Telescope

\u{1f52d}: '🔭'

CrystalBall

\u{1f52e}: '🔮'

SixPointedStarWithMiddleDot

\u{1f52f}: '🔯'

JapaneseSymbolForBeginner

\u{1f530}: '🔰'

TridentEmblem

\u{1f531}: '🔱'

BlackSquareButton

\u{1f532}: '🔲'

WhiteSquareButton

\u{1f533}: '🔳'

LargeRedCircle

\u{1f534}: '🔴'

LargeBlueCircle

\u{1f535}: '🔵'

LargeOrangeDiamond

\u{1f536}: '🔶'

LargeBlueDiamond

\u{1f537}: '🔷'

SmallOrangeDiamond

\u{1f538}: '🔸'

SmallBlueDiamond

\u{1f539}: '🔹'

UpDashPointingRedTriangle

\u{1f53a}: '🔺'

DownDashPointingRedTriangle

\u{1f53b}: '🔻'

UpDashPointingSmallRedTriangle

\u{1f53c}: '🔼'

DownDashPointingSmallRedTriangle

\u{1f53d}: '🔽'

LowerRightShadowedWhiteCircle

\u{1f53e}: '🔾'

UpperRightShadowedWhiteCircle

\u{1f53f}: '🔿'

CircledCrossPommee

\u{1f540}: '🕀'

CrossPommeeWithHalfDashCircleBelow

\u{1f541}: '🕁'

CrossPommee

\u{1f542}: '🕂'

NotchedLeftSemicircleWithThreeDots

\u{1f543}: '🕃'

NotchedRightSemicircleWithThreeDots

\u{1f544}: '🕄'

SymbolForMarksChapter

\u{1f545}: '🕅'

WhiteLatinCross

\u{1f546}: '🕆'

HeavyLatinCross

\u{1f547}: '🕇'

CelticCross

\u{1f548}: '🕈'

OmSymbol

\u{1f549}: '🕉'

DoveOfPeace

\u{1f54a}: '🕊'

Kaaba

\u{1f54b}: '🕋'

Mosque

\u{1f54c}: '🕌'

Synagogue

\u{1f54d}: '🕍'

MenorahWithNineBranches

\u{1f54e}: '🕎'

BowlOfHygieia

\u{1f54f}: '🕏'

ClockFaceOneOclock

\u{1f550}: '🕐'

ClockFaceTwoOclock

\u{1f551}: '🕑'

ClockFaceThreeOclock

\u{1f552}: '🕒'

ClockFaceFourOclock

\u{1f553}: '🕓'

ClockFaceFiveOclock

\u{1f554}: '🕔'

ClockFaceSixOclock

\u{1f555}: '🕕'

ClockFaceSevenOclock

\u{1f556}: '🕖'

ClockFaceEightOclock

\u{1f557}: '🕗'

ClockFaceNineOclock

\u{1f558}: '🕘'

ClockFaceTenOclock

\u{1f559}: '🕙'

ClockFaceElevenOclock

\u{1f55a}: '🕚'

ClockFaceTwelveOclock

\u{1f55b}: '🕛'

ClockFaceOneDashThirty

\u{1f55c}: '🕜'

ClockFaceTwoDashThirty

\u{1f55d}: '🕝'

ClockFaceThreeDashThirty

\u{1f55e}: '🕞'

ClockFaceFourDashThirty

\u{1f55f}: '🕟'

ClockFaceFiveDashThirty

\u{1f560}: '🕠'

ClockFaceSixDashThirty

\u{1f561}: '🕡'

ClockFaceSevenDashThirty

\u{1f562}: '🕢'

ClockFaceEightDashThirty

\u{1f563}: '🕣'

ClockFaceNineDashThirty

\u{1f564}: '🕤'

ClockFaceTenDashThirty

\u{1f565}: '🕥'

ClockFaceElevenDashThirty

\u{1f566}: '🕦'

ClockFaceTwelveDashThirty

\u{1f567}: '🕧'

RightSpeaker

\u{1f568}: '🕨'

RightSpeakerWithOneSoundWave

\u{1f569}: '🕩'

RightSpeakerWithThreeSoundWaves

\u{1f56a}: '🕪'

Bullhorn

\u{1f56b}: '🕫'

BullhornWithSoundWaves

\u{1f56c}: '🕬'

RingingBell

\u{1f56d}: '🕭'

Book

\u{1f56e}: '🕮'

Candle

\u{1f56f}: '🕯'

MantelpieceClock

\u{1f570}: '🕰'

BlackSkullAndCrossbones

\u{1f571}: '🕱'

NoPiracy

\u{1f572}: '🕲'

Hole

\u{1f573}: '🕳'

ManInBusinessSuitLevitating

\u{1f574}: '🕴'

SleuthOrSpy

\u{1f575}: '🕵'

DarkSunglasses

\u{1f576}: '🕶'

Spider

\u{1f577}: '🕷'

SpiderWeb

\u{1f578}: '🕸'

Joystick

\u{1f579}: '🕹'

ManDancing

\u{1f57a}: '🕺'

LeftHandTelephoneReceiver

\u{1f57b}: '🕻'

TelephoneReceiverWithPage

\u{1f57c}: '🕼'

RightHandTelephoneReceiver

\u{1f57d}: '🕽'

WhiteTouchtoneTelephone

\u{1f57e}: '🕾'

BlackTouchtoneTelephone

\u{1f57f}: '🕿'

TelephoneOnTopOfModem

\u{1f580}: '🖀'

ClamshellMobilePhone

\u{1f581}: '🖁'

BackOfEnvelope

\u{1f582}: '🖂'

StampedEnvelope

\u{1f583}: '🖃'

EnvelopeWithLightning

\u{1f584}: '🖄'

FlyingEnvelope

\u{1f585}: '🖅'

PenOverStampedEnvelope

\u{1f586}: '🖆'

LinkedPaperclips

\u{1f587}: '🖇'

BlackPushpin

\u{1f588}: '🖈'

LowerLeftPencil

\u{1f589}: '🖉'

LowerLeftBallpointPen

\u{1f58a}: '🖊'

LowerLeftFountainPen

\u{1f58b}: '🖋'

LowerLeftPaintbrush

\u{1f58c}: '🖌'

LowerLeftCrayon

\u{1f58d}: '🖍'

LeftWritingHand

\u{1f58e}: '🖎'

TurnedOkHandSign

\u{1f58f}: '🖏'

RaisedHandWithFingersSplayed

\u{1f590}: '🖐'

ReversedRaisedHandWithFingersSplayed

\u{1f591}: '🖑'

ReversedThumbsUpSign

\u{1f592}: '🖒'

ReversedThumbsDownSign

\u{1f593}: '🖓'

ReversedVictoryHand

\u{1f594}: '🖔'

ReversedHandWithMiddleFingerExtended

\u{1f595}: '🖕'

RaisedHandWithPartBetweenMiddleAndRingFingers

\u{1f596}: '🖖'

WhiteDownPointingLeftHandIndex

\u{1f597}: '🖗'

SidewaysWhiteLeftPointingIndex

\u{1f598}: '🖘'

SidewaysWhiteRightPointingIndex

\u{1f599}: '🖙'

SidewaysBlackLeftPointingIndex

\u{1f59a}: '🖚'

SidewaysBlackRightPointingIndex

\u{1f59b}: '🖛'

BlackLeftPointingBackhandIndex

\u{1f59c}: '🖜'

BlackRightPointingBackhandIndex

\u{1f59d}: '🖝'

SidewaysWhiteUpPointingIndex

\u{1f59e}: '🖞'

SidewaysWhiteDownPointingIndex

\u{1f59f}: '🖟'

SidewaysBlackUpPointingIndex

\u{1f5a0}: '🖠'

SidewaysBlackDownPointingIndex

\u{1f5a1}: '🖡'

BlackUpPointingBackhandIndex

\u{1f5a2}: '🖢'

BlackDownPointingBackhandIndex

\u{1f5a3}: '🖣'

BlackHeart

\u{1f5a4}: '🖤'

DesktopComputer

\u{1f5a5}: '🖥'

KeyboardAndMouse

\u{1f5a6}: '🖦'

ThreeNetworkedComputers

\u{1f5a7}: '🖧'

Printer

\u{1f5a8}: '🖨'

PocketCalculator

\u{1f5a9}: '🖩'

BlackHardShellFloppyDisk

\u{1f5aa}: '🖪'

WhiteHardShellFloppyDisk

\u{1f5ab}: '🖫'

SoftShellFloppyDisk

\u{1f5ac}: '🖬'

TapeCartridge

\u{1f5ad}: '🖭'

WiredKeyboard

\u{1f5ae}: '🖮'

OneButtonMouse

\u{1f5af}: '🖯'

TwoButtonMouse

\u{1f5b0}: '🖰'

ThreeButtonMouse

\u{1f5b1}: '🖱'

Trackball

\u{1f5b2}: '🖲'

OldPersonalComputer

\u{1f5b3}: '🖳'

HardDisk

\u{1f5b4}: '🖴'

Screen

\u{1f5b5}: '🖵'

PrinterIcon

\u{1f5b6}: '🖶'

FaxIcon

\u{1f5b7}: '🖷'

OpticalDiscIcon

\u{1f5b8}: '🖸'

DocumentWithText

\u{1f5b9}: '🖹'

DocumentWithTextAndPicture

\u{1f5ba}: '🖺'

DocumentWithPicture

\u{1f5bb}: '🖻'

FrameWithPicture

\u{1f5bc}: '🖼'

FrameWithTiles

\u{1f5bd}: '🖽'

FrameWithAnX

\u{1f5be}: '🖾'

BlackFolder

\u{1f5bf}: '🖿'

Folder

\u{1f5c0}: '🗀'

OpenFolder

\u{1f5c1}: '🗁'

CardIndexDividers

\u{1f5c2}: '🗂'

CardFileBox

\u{1f5c3}: '🗃'

FileCabinet

\u{1f5c4}: '🗄'

EmptyNote

\u{1f5c5}: '🗅'

EmptyNotePage

\u{1f5c6}: '🗆'

EmptyNotePad

\u{1f5c7}: '🗇'

Note

\u{1f5c8}: '🗈'

NotePage

\u{1f5c9}: '🗉'

NotePad

\u{1f5ca}: '🗊'

EmptyDocument

\u{1f5cb}: '🗋'

EmptyPage

\u{1f5cc}: '🗌'

EmptyPages

\u{1f5cd}: '🗍'

Document

\u{1f5ce}: '🗎'

Page

\u{1f5cf}: '🗏'

Pages

\u{1f5d0}: '🗐'

Wastebasket

\u{1f5d1}: '🗑'

SpiralNotePad

\u{1f5d2}: '🗒'

SpiralCalendarPad

\u{1f5d3}: '🗓'

DesktopWindow

\u{1f5d4}: '🗔'

Minimize

\u{1f5d5}: '🗕'

Maximize

\u{1f5d6}: '🗖'

Overlap

\u{1f5d7}: '🗗'

ClockwiseRightAndLeftSemicircleArrows

\u{1f5d8}: '🗘'

CancellationX

\u{1f5d9}: '🗙'

IncreaseFontSizeSymbol

\u{1f5da}: '🗚'

DecreaseFontSizeSymbol

\u{1f5db}: '🗛'

Compression

\u{1f5dc}: '🗜'

OldKey

\u{1f5dd}: '🗝'

RolledDashUpNewspaper

\u{1f5de}: '🗞'

PageWithCircledText

\u{1f5df}: '🗟'

StockChart

\u{1f5e0}: '🗠'

DaggerKnife

\u{1f5e1}: '🗡'

Lips

\u{1f5e2}: '🗢'

SpeakingHeadInSilhouette

\u{1f5e3}: '🗣'

ThreeRaysAbove

\u{1f5e4}: '🗤'

ThreeRaysBelow

\u{1f5e5}: '🗥'

ThreeRaysLeft

\u{1f5e6}: '🗦'

ThreeRaysRight

\u{1f5e7}: '🗧'

LeftSpeechBubble

\u{1f5e8}: '🗨'

RightSpeechBubble

\u{1f5e9}: '🗩'

TwoSpeechBubbles

\u{1f5ea}: '🗪'

ThreeSpeechBubbles

\u{1f5eb}: '🗫'

LeftThoughtBubble

\u{1f5ec}: '🗬'

RightThoughtBubble

\u{1f5ed}: '🗭'

LeftAngerBubble

\u{1f5ee}: '🗮'

RightAngerBubble

\u{1f5ef}: '🗯'

MoodBubble

\u{1f5f0}: '🗰'

LightningMoodBubble

\u{1f5f1}: '🗱'

LightningMood

\u{1f5f2}: '🗲'

BallotBoxWithBallot

\u{1f5f3}: '🗳'

BallotScriptX

\u{1f5f4}: '🗴'

BallotBoxWithScriptX

\u{1f5f5}: '🗵'

BallotBoldScriptX

\u{1f5f6}: '🗶'

BallotBoxWithBoldScriptX

\u{1f5f7}: '🗷'

LightCheckMark

\u{1f5f8}: '🗸'

BallotBoxWithBoldCheck

\u{1f5f9}: '🗹'

WorldMap

\u{1f5fa}: '🗺'

MountFuji

\u{1f5fb}: '🗻'

TokyoTower

\u{1f5fc}: '🗼'

StatueOfLiberty

\u{1f5fd}: '🗽'

SilhouetteOfJapan

\u{1f5fe}: '🗾'

Methods

impl MiscellaneousSymbolsandPictographs

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case