Enum unicode_types::miscellaneous_technical::MiscellaneousTechnical

pub enum MiscellaneousTechnical {
    DiameterSign,
    ElectricArrow,
    House,
    UpArrowhead,
    DownArrowhead,
    Projective,
    Perspective,
    WavyLine,
    LeftCeiling,
    RightCeiling,
    LeftFloor,
    RightFloor,
    BottomRightCrop,
    BottomLeftCrop,
    TopRightCrop,
    TopLeftCrop,
    ReversedNotSign,
    SquareLozenge,
    Arc,
    Segment,
    Sector,
    TelephoneRecorder,
    PositionIndicator,
    ViewdataSquare,
    PlaceOfInterestSign,
    TurnedNotSign,
    Watch,
    Hourglass,
    TopLeftCorner,
    TopRightCorner,
    BottomLeftCorner,
    BottomRightCorner,
    TopHalfIntegral,
    BottomHalfIntegral,
    Frown,
    Smile,
    UpArrowheadBetweenTwoHorizontalBars,
    OptionKey,
    EraseToTheRight,
    XInARectangleBox,
    Keyboard,
    LeftDashPointingAngleBracket,
    RightDashPointingAngleBracket,
    EraseToTheLeft,
    BenzeneRing,
    Cylindricity,
    AllAroundDashProfile,
    Symmetry,
    TotalRunout,
    DimensionOrigin,
    ConicalTaper,
    Slope,
    Counterbore,
    Countersink,
    AplFunctionalSymbolIDashBeam,
    AplFunctionalSymbolSquishQuad,
    AplFunctionalSymbolQuadEqual,
    AplFunctionalSymbolQuadDivide,
    AplFunctionalSymbolQuadDiamond,
    AplFunctionalSymbolQuadJot,
    AplFunctionalSymbolQuadCircle,
    AplFunctionalSymbolCircleStile,
    AplFunctionalSymbolCircleJot,
    AplFunctionalSymbolSlashBar,
    AplFunctionalSymbolBackslashBar,
    AplFunctionalSymbolQuadSlash,
    AplFunctionalSymbolQuadBackslash,
    AplFunctionalSymbolQuadLessDashThan,
    AplFunctionalSymbolQuadGreaterDashThan,
    AplFunctionalSymbolLeftwardsVane,
    AplFunctionalSymbolRightwardsVane,
    AplFunctionalSymbolQuadLeftwardsArrow,
    AplFunctionalSymbolQuadRightwardsArrow,
    AplFunctionalSymbolCircleBackslash,
    AplFunctionalSymbolDownTackUnderbar,
    AplFunctionalSymbolDeltaStile,
    AplFunctionalSymbolQuadDownCaret,
    AplFunctionalSymbolQuadDelta,
    AplFunctionalSymbolDownTackJot,
    AplFunctionalSymbolUpwardsVane,
    AplFunctionalSymbolQuadUpwardsArrow,
    AplFunctionalSymbolUpTackOverbar,
    AplFunctionalSymbolDelStile,
    AplFunctionalSymbolQuadUpCaret,
    AplFunctionalSymbolQuadDel,
    AplFunctionalSymbolUpTackJot,
    AplFunctionalSymbolDownwardsVane,
    AplFunctionalSymbolQuadDownwardsArrow,
    AplFunctionalSymbolQuoteUnderbar,
    AplFunctionalSymbolDeltaUnderbar,
    AplFunctionalSymbolDiamondUnderbar,
    AplFunctionalSymbolJotUnderbar,
    AplFunctionalSymbolCircleUnderbar,
    AplFunctionalSymbolUpShoeJot,
    AplFunctionalSymbolQuoteQuad,
    AplFunctionalSymbolCircleStar,
    AplFunctionalSymbolQuadColon,
    AplFunctionalSymbolUpTackDiaeresis,
    AplFunctionalSymbolDelDiaeresis,
    AplFunctionalSymbolStarDiaeresis,
    AplFunctionalSymbolJotDiaeresis,
    AplFunctionalSymbolCircleDiaeresis,
    AplFunctionalSymbolDownShoeStile,
    AplFunctionalSymbolLeftShoeStile,
    AplFunctionalSymbolTildeDiaeresis,
    AplFunctionalSymbolGreaterDashThanDiaeresis,
    AplFunctionalSymbolCommaBar,
    AplFunctionalSymbolDelTilde,
    AplFunctionalSymbolZilde,
    AplFunctionalSymbolStileTilde,
    AplFunctionalSymbolSemicolonUnderbar,
    AplFunctionalSymbolQuadNotEqual,
    AplFunctionalSymbolQuadQuestion,
    AplFunctionalSymbolDownCaretTilde,
    AplFunctionalSymbolUpCaretTilde,
    AplFunctionalSymbolIota,
    AplFunctionalSymbolRho,
    AplFunctionalSymbolOmega,
    AplFunctionalSymbolAlphaUnderbar,
    AplFunctionalSymbolEpsilonUnderbar,
    AplFunctionalSymbolIotaUnderbar,
    AplFunctionalSymbolOmegaUnderbar,
    AplFunctionalSymbolAlpha,
    NotCheckMark,
    RightAngleWithDownwardsZigzagArrow,
    ShoulderedOpenBox,
    BellSymbol,
    VerticalLineWithMiddleDot,
    InsertionSymbol,
    ContinuousUnderlineSymbol,
    DiscontinuousUnderlineSymbol,
    EmphasisSymbol,
    CompositionSymbol,
    WhiteSquareWithCentreVerticalLine,
    EnterSymbol,
    AlternativeKeySymbol,
    HelmSymbol,
    CircledHorizontalBarWithNotch,
    CircledTriangleDown,
    BrokenCircleWithNorthwestArrow,
    UndoSymbol,
    MonostableSymbol,
    HysteresisSymbol,
    OpenDashCircuitDashOutputHDashTypeSymbol,
    OpenDashCircuitDashOutputLDashTypeSymbol,
    PassiveDashPullDashDownDashOutputSymbol,
    PassiveDashPullDashUpDashOutputSymbol,
    DirectCurrentSymbolFormTwo,
    SoftwareDashFunctionSymbol,
    AplFunctionalSymbolQuad,
    DecimalSeparatorKeySymbol,
    PreviousPage,
    NextPage,
    PrintScreenSymbol,
    ClearScreenSymbol,
    LeftParenthesisUpperHook,
    LeftParenthesisExtension,
    LeftParenthesisLowerHook,
    RightParenthesisUpperHook,
    RightParenthesisExtension,
    RightParenthesisLowerHook,
    LeftSquareBracketUpperCorner,
    LeftSquareBracketExtension,
    LeftSquareBracketLowerCorner,
    RightSquareBracketUpperCorner,
    RightSquareBracketExtension,
    RightSquareBracketLowerCorner,
    LeftCurlyBracketUpperHook,
    LeftCurlyBracketMiddlePiece,
    LeftCurlyBracketLowerHook,
    CurlyBracketExtension,
    RightCurlyBracketUpperHook,
    RightCurlyBracketMiddlePiece,
    RightCurlyBracketLowerHook,
    IntegralExtension,
    HorizontalLineExtension,
    UpperLeftOrLowerRightCurlyBracketSection,
    UpperRightOrLowerLeftCurlyBracketSection,
    SummationTop,
    SummationBottom,
    TopSquareBracket,
    BottomSquareBracket,
    BottomSquareBracketOverTopSquareBracket,
    RadicalSymbolBottom,
    LeftVerticalBoxLine,
    RightVerticalBoxLine,
    HorizontalScanLineDash1,
    HorizontalScanLineDash3,
    HorizontalScanLineDash7,
    HorizontalScanLineDash9,
    DentistrySymbolLightVerticalAndTopRight,
    DentistrySymbolLightVerticalAndBottomRight,
    DentistrySymbolLightVerticalWithCircle,
    DentistrySymbolLightDownAndHorizontalWithCircle,
    DentistrySymbolLightUpAndHorizontalWithCircle,
    DentistrySymbolLightVerticalWithTriangle,
    DentistrySymbolLightDownAndHorizontalWithTriangle,
    DentistrySymbolLightUpAndHorizontalWithTriangle,
    DentistrySymbolLightVerticalAndWave,
    DentistrySymbolLightDownAndHorizontalWithWave,
    DentistrySymbolLightUpAndHorizontalWithWave,
    DentistrySymbolLightDownAndHorizontal,
    DentistrySymbolLightUpAndHorizontal,
    DentistrySymbolLightVerticalAndTopLeft,
    DentistrySymbolLightVerticalAndBottomLeft,
    SquareFoot,
    ReturnSymbol,
    EjectSymbol,
    VerticalLineExtension,
    MetricalBreve,
    MetricalLongOverShort,
    MetricalShortOverLong,
    MetricalLongOverTwoShorts,
    MetricalTwoShortsOverLong,
    MetricalTwoShortsJoined,
    MetricalTriseme,
    MetricalTetraseme,
    MetricalPentaseme,
    EarthGround,
    Fuse,
    TopParenthesis,
    BottomParenthesis,
    TopCurlyBracket,
    BottomCurlyBracket,
    TopTortoiseShellBracket,
    BottomTortoiseShellBracket,
    WhiteTrapezium,
    BenzeneRingWithCircle,
    Straightness,
    Flatness,
    AcCurrent,
    ElectricalIntersection,
    DecimalExponentSymbol,
    BlackRightDashPointingDoubleTriangle,
    BlackLeftDashPointingDoubleTriangle,
    BlackUpDashPointingDoubleTriangle,
    BlackDownDashPointingDoubleTriangle,
    BlackRightDashPointingDoubleTriangleWithVerticalBar,
    BlackLeftDashPointingDoubleTriangleWithVerticalBar,
    BlackRightDashPointingTriangleWithDoubleVerticalBar,
    AlarmClock,
    Stopwatch,
    TimerClock,
    HourglassWithFlowingSand,
    BlackMediumLeftDashPointingTriangle,
    BlackMediumRightDashPointingTriangle,
    BlackMediumUpDashPointingTriangle,
    BlackMediumDownDashPointingTriangle,
    DoubleVerticalBar,
    BlackSquareForStop,
    BlackCircleForRecord,
    PowerSymbol,
    PowerOnDashOffSymbol,
    PowerOnSymbol,
    PowerSleepSymbol,
}

An enum to represent all characters in the MiscellaneousTechnical block.

Variants
DiameterSign

\u{2300}: '⌀'

ElectricArrow

\u{2301}: '⌁'

House

\u{2302}: '⌂'

UpArrowhead

\u{2303}: '⌃'

DownArrowhead

\u{2304}: '⌄'

Projective

\u{2305}: '⌅'

Perspective

\u{2306}: '⌆'

WavyLine

\u{2307}: '⌇'

LeftCeiling

\u{2308}: '⌈'

RightCeiling

\u{2309}: '⌉'

LeftFloor

\u{230a}: '⌊'

RightFloor

\u{230b}: '⌋'

BottomRightCrop

\u{230c}: '⌌'

BottomLeftCrop

\u{230d}: '⌍'

TopRightCrop

\u{230e}: '⌎'

TopLeftCrop

\u{230f}: '⌏'

ReversedNotSign

\u{2310}: '⌐'

SquareLozenge

\u{2311}: '⌑'

Arc

\u{2312}: '⌒'

Segment

\u{2313}: '⌓'

Sector

\u{2314}: '⌔'

TelephoneRecorder

\u{2315}: '⌕'

PositionIndicator

\u{2316}: '⌖'

ViewdataSquare

\u{2317}: '⌗'

PlaceOfInterestSign

\u{2318}: '⌘'

TurnedNotSign

\u{2319}: '⌙'

Watch

\u{231a}: '⌚'

Hourglass

\u{231b}: '⌛'

TopLeftCorner

\u{231c}: '⌜'

TopRightCorner

\u{231d}: '⌝'

BottomLeftCorner

\u{231e}: '⌞'

BottomRightCorner

\u{231f}: '⌟'

TopHalfIntegral

\u{2320}: '⌠'

BottomHalfIntegral

\u{2321}: '⌡'

Frown

\u{2322}: '⌢'

Smile

\u{2323}: '⌣'

UpArrowheadBetweenTwoHorizontalBars

\u{2324}: '⌤'

OptionKey

\u{2325}: '⌥'

EraseToTheRight

\u{2326}: '⌦'

XInARectangleBox

\u{2327}: '⌧'

Keyboard

\u{2328}: '⌨'

LeftDashPointingAngleBracket

\u{2329}: '〈'

RightDashPointingAngleBracket

\u{232a}: '〉'

EraseToTheLeft

\u{232b}: '⌫'

BenzeneRing

\u{232c}: '⌬'

Cylindricity

\u{232d}: '⌭'

AllAroundDashProfile

\u{232e}: '⌮'

Symmetry

\u{232f}: '⌯'

TotalRunout

\u{2330}: '⌰'

DimensionOrigin

\u{2331}: '⌱'

ConicalTaper

\u{2332}: '⌲'

Slope

\u{2333}: '⌳'

Counterbore

\u{2334}: '⌴'

Countersink

\u{2335}: '⌵'

AplFunctionalSymbolIDashBeam

\u{2336}: '⌶'

AplFunctionalSymbolSquishQuad

\u{2337}: '⌷'

AplFunctionalSymbolQuadEqual

\u{2338}: '⌸'

AplFunctionalSymbolQuadDivide

\u{2339}: '⌹'

AplFunctionalSymbolQuadDiamond

\u{233a}: '⌺'

AplFunctionalSymbolQuadJot

\u{233b}: '⌻'

AplFunctionalSymbolQuadCircle

\u{233c}: '⌼'

AplFunctionalSymbolCircleStile

\u{233d}: '⌽'

AplFunctionalSymbolCircleJot

\u{233e}: '⌾'

AplFunctionalSymbolSlashBar

\u{233f}: '⌿'

AplFunctionalSymbolBackslashBar

\u{2340}: '⍀'

AplFunctionalSymbolQuadSlash

\u{2341}: '⍁'

AplFunctionalSymbolQuadBackslash

\u{2342}: '⍂'

AplFunctionalSymbolQuadLessDashThan

\u{2343}: '⍃'

AplFunctionalSymbolQuadGreaterDashThan

\u{2344}: '⍄'

AplFunctionalSymbolLeftwardsVane

\u{2345}: '⍅'

AplFunctionalSymbolRightwardsVane

\u{2346}: '⍆'

AplFunctionalSymbolQuadLeftwardsArrow

\u{2347}: '⍇'

AplFunctionalSymbolQuadRightwardsArrow

\u{2348}: '⍈'

AplFunctionalSymbolCircleBackslash

\u{2349}: '⍉'

AplFunctionalSymbolDownTackUnderbar

\u{234a}: '⍊'

AplFunctionalSymbolDeltaStile

\u{234b}: '⍋'

AplFunctionalSymbolQuadDownCaret

\u{234c}: '⍌'

AplFunctionalSymbolQuadDelta

\u{234d}: '⍍'

AplFunctionalSymbolDownTackJot

\u{234e}: '⍎'

AplFunctionalSymbolUpwardsVane

\u{234f}: '⍏'

AplFunctionalSymbolQuadUpwardsArrow

\u{2350}: '⍐'

AplFunctionalSymbolUpTackOverbar

\u{2351}: '⍑'

AplFunctionalSymbolDelStile

\u{2352}: '⍒'

AplFunctionalSymbolQuadUpCaret

\u{2353}: '⍓'

AplFunctionalSymbolQuadDel

\u{2354}: '⍔'

AplFunctionalSymbolUpTackJot

\u{2355}: '⍕'

AplFunctionalSymbolDownwardsVane

\u{2356}: '⍖'

AplFunctionalSymbolQuadDownwardsArrow

\u{2357}: '⍗'

AplFunctionalSymbolQuoteUnderbar

\u{2358}: '⍘'

AplFunctionalSymbolDeltaUnderbar

\u{2359}: '⍙'

AplFunctionalSymbolDiamondUnderbar

\u{235a}: '⍚'

AplFunctionalSymbolJotUnderbar

\u{235b}: '⍛'

AplFunctionalSymbolCircleUnderbar

\u{235c}: '⍜'

AplFunctionalSymbolUpShoeJot

\u{235d}: '⍝'

AplFunctionalSymbolQuoteQuad

\u{235e}: '⍞'

AplFunctionalSymbolCircleStar

\u{235f}: '⍟'

AplFunctionalSymbolQuadColon

\u{2360}: '⍠'

AplFunctionalSymbolUpTackDiaeresis

\u{2361}: '⍡'

AplFunctionalSymbolDelDiaeresis

\u{2362}: '⍢'

AplFunctionalSymbolStarDiaeresis

\u{2363}: '⍣'

AplFunctionalSymbolJotDiaeresis

\u{2364}: '⍤'

AplFunctionalSymbolCircleDiaeresis

\u{2365}: '⍥'

AplFunctionalSymbolDownShoeStile

\u{2366}: '⍦'

AplFunctionalSymbolLeftShoeStile

\u{2367}: '⍧'

AplFunctionalSymbolTildeDiaeresis

\u{2368}: '⍨'

AplFunctionalSymbolGreaterDashThanDiaeresis

\u{2369}: '⍩'

AplFunctionalSymbolCommaBar

\u{236a}: '⍪'

AplFunctionalSymbolDelTilde

\u{236b}: '⍫'

AplFunctionalSymbolZilde

\u{236c}: '⍬'

AplFunctionalSymbolStileTilde

\u{236d}: '⍭'

AplFunctionalSymbolSemicolonUnderbar

\u{236e}: '⍮'

AplFunctionalSymbolQuadNotEqual

\u{236f}: '⍯'

AplFunctionalSymbolQuadQuestion

\u{2370}: '⍰'

AplFunctionalSymbolDownCaretTilde

\u{2371}: '⍱'

AplFunctionalSymbolUpCaretTilde

\u{2372}: '⍲'

AplFunctionalSymbolIota

\u{2373}: '⍳'

AplFunctionalSymbolRho

\u{2374}: '⍴'

AplFunctionalSymbolOmega

\u{2375}: '⍵'

AplFunctionalSymbolAlphaUnderbar

\u{2376}: '⍶'

AplFunctionalSymbolEpsilonUnderbar

\u{2377}: '⍷'

AplFunctionalSymbolIotaUnderbar

\u{2378}: '⍸'

AplFunctionalSymbolOmegaUnderbar

\u{2379}: '⍹'

AplFunctionalSymbolAlpha

\u{237a}: '⍺'

NotCheckMark

\u{237b}: '⍻'

RightAngleWithDownwardsZigzagArrow

\u{237c}: '⍼'

ShoulderedOpenBox

\u{237d}: '⍽'

BellSymbol

\u{237e}: '⍾'

VerticalLineWithMiddleDot

\u{237f}: '⍿'

InsertionSymbol

\u{2380}: '⎀'

ContinuousUnderlineSymbol

\u{2381}: '⎁'

DiscontinuousUnderlineSymbol

\u{2382}: '⎂'

EmphasisSymbol

\u{2383}: '⎃'

CompositionSymbol

\u{2384}: '⎄'

WhiteSquareWithCentreVerticalLine

\u{2385}: '⎅'

EnterSymbol

\u{2386}: '⎆'

AlternativeKeySymbol

\u{2387}: '⎇'

HelmSymbol

\u{2388}: '⎈'

CircledHorizontalBarWithNotch

\u{2389}: '⎉'

CircledTriangleDown

\u{238a}: '⎊'

BrokenCircleWithNorthwestArrow

\u{238b}: '⎋'

UndoSymbol

\u{238c}: '⎌'

MonostableSymbol

\u{238d}: '⎍'

HysteresisSymbol

\u{238e}: '⎎'

OpenDashCircuitDashOutputHDashTypeSymbol

\u{238f}: '⎏'

OpenDashCircuitDashOutputLDashTypeSymbol

\u{2390}: '⎐'

PassiveDashPullDashDownDashOutputSymbol

\u{2391}: '⎑'

PassiveDashPullDashUpDashOutputSymbol

\u{2392}: '⎒'

DirectCurrentSymbolFormTwo

\u{2393}: '⎓'

SoftwareDashFunctionSymbol

\u{2394}: '⎔'

AplFunctionalSymbolQuad

\u{2395}: '⎕'

DecimalSeparatorKeySymbol

\u{2396}: '⎖'

PreviousPage

\u{2397}: '⎗'

NextPage

\u{2398}: '⎘'

PrintScreenSymbol

\u{2399}: '⎙'

ClearScreenSymbol

\u{239a}: '⎚'

LeftParenthesisUpperHook

\u{239b}: '⎛'

LeftParenthesisExtension

\u{239c}: '⎜'

LeftParenthesisLowerHook

\u{239d}: '⎝'

RightParenthesisUpperHook

\u{239e}: '⎞'

RightParenthesisExtension

\u{239f}: '⎟'

RightParenthesisLowerHook

\u{23a0}: '⎠'

LeftSquareBracketUpperCorner

\u{23a1}: '⎡'

LeftSquareBracketExtension

\u{23a2}: '⎢'

LeftSquareBracketLowerCorner

\u{23a3}: '⎣'

RightSquareBracketUpperCorner

\u{23a4}: '⎤'

RightSquareBracketExtension

\u{23a5}: '⎥'

RightSquareBracketLowerCorner

\u{23a6}: '⎦'

LeftCurlyBracketUpperHook

\u{23a7}: '⎧'

LeftCurlyBracketMiddlePiece

\u{23a8}: '⎨'

LeftCurlyBracketLowerHook

\u{23a9}: '⎩'

CurlyBracketExtension

\u{23aa}: '⎪'

RightCurlyBracketUpperHook

\u{23ab}: '⎫'

RightCurlyBracketMiddlePiece

\u{23ac}: '⎬'

RightCurlyBracketLowerHook

\u{23ad}: '⎭'

IntegralExtension

\u{23ae}: '⎮'

HorizontalLineExtension

\u{23af}: '⎯'

UpperLeftOrLowerRightCurlyBracketSection

\u{23b0}: '⎰'

UpperRightOrLowerLeftCurlyBracketSection

\u{23b1}: '⎱'

SummationTop

\u{23b2}: '⎲'

SummationBottom

\u{23b3}: '⎳'

TopSquareBracket

\u{23b4}: '⎴'

BottomSquareBracket

\u{23b5}: '⎵'

BottomSquareBracketOverTopSquareBracket

\u{23b6}: '⎶'

RadicalSymbolBottom

\u{23b7}: '⎷'

LeftVerticalBoxLine

\u{23b8}: '⎸'

RightVerticalBoxLine

\u{23b9}: '⎹'

HorizontalScanLineDash1

\u{23ba}: '⎺'

HorizontalScanLineDash3

\u{23bb}: '⎻'

HorizontalScanLineDash7

\u{23bc}: '⎼'

HorizontalScanLineDash9

\u{23bd}: '⎽'

DentistrySymbolLightVerticalAndTopRight

\u{23be}: '⎾'

DentistrySymbolLightVerticalAndBottomRight

\u{23bf}: '⎿'

DentistrySymbolLightVerticalWithCircle

\u{23c0}: '⏀'

DentistrySymbolLightDownAndHorizontalWithCircle

\u{23c1}: '⏁'

DentistrySymbolLightUpAndHorizontalWithCircle

\u{23c2}: '⏂'

DentistrySymbolLightVerticalWithTriangle

\u{23c3}: '⏃'

DentistrySymbolLightDownAndHorizontalWithTriangle

\u{23c4}: '⏄'

DentistrySymbolLightUpAndHorizontalWithTriangle

\u{23c5}: '⏅'

DentistrySymbolLightVerticalAndWave

\u{23c6}: '⏆'

DentistrySymbolLightDownAndHorizontalWithWave

\u{23c7}: '⏇'

DentistrySymbolLightUpAndHorizontalWithWave

\u{23c8}: '⏈'

DentistrySymbolLightDownAndHorizontal

\u{23c9}: '⏉'

DentistrySymbolLightUpAndHorizontal

\u{23ca}: '⏊'

DentistrySymbolLightVerticalAndTopLeft

\u{23cb}: '⏋'

DentistrySymbolLightVerticalAndBottomLeft

\u{23cc}: '⏌'

SquareFoot

\u{23cd}: '⏍'

ReturnSymbol

\u{23ce}: '⏎'

EjectSymbol

\u{23cf}: '⏏'

VerticalLineExtension

\u{23d0}: '⏐'

MetricalBreve

\u{23d1}: '⏑'

MetricalLongOverShort

\u{23d2}: '⏒'

MetricalShortOverLong

\u{23d3}: '⏓'

MetricalLongOverTwoShorts

\u{23d4}: '⏔'

MetricalTwoShortsOverLong

\u{23d5}: '⏕'

MetricalTwoShortsJoined

\u{23d6}: '⏖'

MetricalTriseme

\u{23d7}: '⏗'

MetricalTetraseme

\u{23d8}: '⏘'

MetricalPentaseme

\u{23d9}: '⏙'

EarthGround

\u{23da}: '⏚'

Fuse

\u{23db}: '⏛'

TopParenthesis

\u{23dc}: '⏜'

BottomParenthesis

\u{23dd}: '⏝'

TopCurlyBracket

\u{23de}: '⏞'

BottomCurlyBracket

\u{23df}: '⏟'

TopTortoiseShellBracket

\u{23e0}: '⏠'

BottomTortoiseShellBracket

\u{23e1}: '⏡'

WhiteTrapezium

\u{23e2}: '⏢'

BenzeneRingWithCircle

\u{23e3}: '⏣'

Straightness

\u{23e4}: '⏤'

Flatness

\u{23e5}: '⏥'

AcCurrent

\u{23e6}: '⏦'

ElectricalIntersection

\u{23e7}: '⏧'

DecimalExponentSymbol

\u{23e8}: '⏨'

BlackRightDashPointingDoubleTriangle

\u{23e9}: '⏩'

BlackLeftDashPointingDoubleTriangle

\u{23ea}: '⏪'

BlackUpDashPointingDoubleTriangle

\u{23eb}: '⏫'

BlackDownDashPointingDoubleTriangle

\u{23ec}: '⏬'

BlackRightDashPointingDoubleTriangleWithVerticalBar

\u{23ed}: '⏭'

BlackLeftDashPointingDoubleTriangleWithVerticalBar

\u{23ee}: '⏮'

BlackRightDashPointingTriangleWithDoubleVerticalBar

\u{23ef}: '⏯'

AlarmClock

\u{23f0}: '⏰'

Stopwatch

\u{23f1}: '⏱'

TimerClock

\u{23f2}: '⏲'

HourglassWithFlowingSand

\u{23f3}: '⏳'

BlackMediumLeftDashPointingTriangle

\u{23f4}: '⏴'

BlackMediumRightDashPointingTriangle

\u{23f5}: '⏵'

BlackMediumUpDashPointingTriangle

\u{23f6}: '⏶'

BlackMediumDownDashPointingTriangle

\u{23f7}: '⏷'

DoubleVerticalBar

\u{23f8}: '⏸'

BlackSquareForStop

\u{23f9}: '⏹'

BlackCircleForRecord

\u{23fa}: '⏺'

PowerSymbol

\u{23fb}: '⏻'

PowerOnDashOffSymbol

\u{23fc}: '⏼'

PowerOnSymbol

\u{23fd}: '⏽'

PowerSleepSymbol

\u{23fe}: '⏾'

Methods

impl MiscellaneousTechnical

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case