Enum unicode_types::musical_symbols::MusicalSymbols

pub enum MusicalSymbols {
    MusicalSymbolSingleBarline,
    MusicalSymbolDoubleBarline,
    MusicalSymbolFinalBarline,
    MusicalSymbolReverseFinalBarline,
    MusicalSymbolDashedBarline,
    MusicalSymbolShortBarline,
    MusicalSymbolLeftRepeatSign,
    MusicalSymbolRightRepeatSign,
    MusicalSymbolRepeatDots,
    MusicalSymbolDalSegno,
    MusicalSymbolDaCapo,
    MusicalSymbolSegno,
    MusicalSymbolCoda,
    MusicalSymbolRepeatedFigureDash1,
    MusicalSymbolRepeatedFigureDash2,
    MusicalSymbolRepeatedFigureDash3,
    MusicalSymbolFermata,
    MusicalSymbolFermataBelow,
    MusicalSymbolBreathMark,
    MusicalSymbolCaesura,
    MusicalSymbolBrace,
    MusicalSymbolBracket,
    MusicalSymbolOneDashLineStaff,
    MusicalSymbolTwoDashLineStaff,
    MusicalSymbolThreeDashLineStaff,
    MusicalSymbolFourDashLineStaff,
    MusicalSymbolFiveDashLineStaff,
    MusicalSymbolSixDashLineStaff,
    MusicalSymbolSixDashStringFretboard,
    MusicalSymbolFourDashStringFretboard,
    MusicalSymbolGClef,
    MusicalSymbolGClefOttavaAlta,
    MusicalSymbolGClefOttavaBassa,
    MusicalSymbolCClef,
    MusicalSymbolFClef,
    MusicalSymbolFClefOttavaAlta,
    MusicalSymbolFClefOttavaBassa,
    MusicalSymbolDrumClefDash1,
    MusicalSymbolDrumClefDash2,
    MusicalSymbolMultipleMeasureRest,
    MusicalSymbolDoubleSharp,
    MusicalSymbolDoubleFlat,
    MusicalSymbolFlatUp,
    MusicalSymbolFlatDown,
    MusicalSymbolNaturalUp,
    MusicalSymbolNaturalDown,
    MusicalSymbolSharpUp,
    MusicalSymbolSharpDown,
    MusicalSymbolQuarterToneSharp,
    MusicalSymbolQuarterToneFlat,
    MusicalSymbolCommonTime,
    MusicalSymbolCutTime,
    MusicalSymbolOttavaAlta,
    MusicalSymbolOttavaBassa,
    MusicalSymbolQuindicesimaAlta,
    MusicalSymbolQuindicesimaBassa,
    MusicalSymbolMultiRest,
    MusicalSymbolWholeRest,
    MusicalSymbolHalfRest,
    MusicalSymbolQuarterRest,
    MusicalSymbolEighthRest,
    MusicalSymbolSixteenthRest,
    MusicalSymbolThirtyDashSecondRest,
    MusicalSymbolSixtyDashFourthRest,
    MusicalSymbolOneHundredTwentyDashEighthRest,
    MusicalSymbolXNotehead,
    MusicalSymbolPlusNotehead,
    MusicalSymbolCircleXNotehead,
    MusicalSymbolSquareNoteheadWhite,
    MusicalSymbolSquareNoteheadBlack,
    MusicalSymbolTriangleNoteheadUpWhite,
    MusicalSymbolTriangleNoteheadUpBlack,
    MusicalSymbolTriangleNoteheadLeftWhite,
    MusicalSymbolTriangleNoteheadLeftBlack,
    MusicalSymbolTriangleNoteheadRightWhite,
    MusicalSymbolTriangleNoteheadRightBlack,
    MusicalSymbolTriangleNoteheadDownWhite,
    MusicalSymbolTriangleNoteheadDownBlack,
    MusicalSymbolTriangleNoteheadUpRightWhite,
    MusicalSymbolTriangleNoteheadUpRightBlack,
    MusicalSymbolMoonNoteheadWhite,
    MusicalSymbolMoonNoteheadBlack,
    MusicalSymbolTriangleDashRoundNoteheadDownWhite,
    MusicalSymbolTriangleDashRoundNoteheadDownBlack,
    MusicalSymbolParenthesisNotehead,
    MusicalSymbolVoidNotehead,
    MusicalSymbolNoteheadBlack,
    MusicalSymbolNullNotehead,
    MusicalSymbolClusterNoteheadWhite,
    MusicalSymbolClusterNoteheadBlack,
    MusicalSymbolBreve,
    MusicalSymbolWholeNote,
    MusicalSymbolHalfNote,
    MusicalSymbolQuarterNote,
    MusicalSymbolEighthNote,
    MusicalSymbolSixteenthNote,
    MusicalSymbolThirtyDashSecondNote,
    MusicalSymbolSixtyDashFourthNote,
    MusicalSymbolOneHundredTwentyDashEighthNote,
    MusicalSymbolCombiningStem,
    MusicalSymbolCombiningSprechgesangStem,
    MusicalSymbolCombiningTremoloDash1,
    MusicalSymbolCombiningTremoloDash2,
    MusicalSymbolCombiningTremoloDash3,
    MusicalSymbolFingeredTremoloDash1,
    MusicalSymbolFingeredTremoloDash2,
    MusicalSymbolFingeredTremoloDash3,
    MusicalSymbolCombiningAugmentationDot,
    MusicalSymbolCombiningFlagDash1,
    MusicalSymbolCombiningFlagDash2,
    MusicalSymbolCombiningFlagDash3,
    MusicalSymbolCombiningFlagDash4,
    MusicalSymbolCombiningFlagDash5,
    MusicalSymbolBeginBeam,
    MusicalSymbolEndBeam,
    MusicalSymbolBeginTie,
    MusicalSymbolEndTie,
    MusicalSymbolBeginSlur,
    MusicalSymbolEndSlur,
    MusicalSymbolBeginPhrase,
    MusicalSymbolEndPhrase,
    MusicalSymbolCombiningAccent,
    MusicalSymbolCombiningStaccato,
    MusicalSymbolCombiningTenuto,
    MusicalSymbolCombiningStaccatissimo,
    MusicalSymbolCombiningMarcato,
    MusicalSymbolCombiningMarcatoDashStaccato,
    MusicalSymbolCombiningAccentDashStaccato,
    MusicalSymbolCombiningLoure,
    MusicalSymbolArpeggiatoUp,
    MusicalSymbolArpeggiatoDown,
    MusicalSymbolCombiningDoit,
    MusicalSymbolCombiningRip,
    MusicalSymbolCombiningFlip,
    MusicalSymbolCombiningSmear,
    MusicalSymbolCombiningBend,
    MusicalSymbolCombiningDoubleTongue,
    MusicalSymbolCombiningTripleTongue,
    MusicalSymbolRinforzando,
    MusicalSymbolSubito,
    MusicalSymbolZ,
    MusicalSymbolPiano,
    MusicalSymbolMezzo,
    MusicalSymbolForte,
    MusicalSymbolCrescendo,
    MusicalSymbolDecrescendo,
    MusicalSymbolGraceNoteSlash,
    MusicalSymbolGraceNoteNoSlash,
    MusicalSymbolTr,
    MusicalSymbolTurn,
    MusicalSymbolInvertedTurn,
    MusicalSymbolTurnSlash,
    MusicalSymbolTurnUp,
    MusicalSymbolOrnamentStrokeDash1,
    MusicalSymbolOrnamentStrokeDash2,
    MusicalSymbolOrnamentStrokeDash3,
    MusicalSymbolOrnamentStrokeDash4,
    MusicalSymbolOrnamentStrokeDash5,
    MusicalSymbolOrnamentStrokeDash6,
    MusicalSymbolOrnamentStrokeDash7,
    MusicalSymbolOrnamentStrokeDash8,
    MusicalSymbolOrnamentStrokeDash9,
    MusicalSymbolOrnamentStrokeDash10,
    MusicalSymbolOrnamentStrokeDash11,
    MusicalSymbolHauptstimme,
    MusicalSymbolNebenstimme,
    MusicalSymbolEndOfStimme,
    MusicalSymbolDegreeSlash,
    MusicalSymbolCombiningDownBow,
    MusicalSymbolCombiningUpBow,
    MusicalSymbolCombiningHarmonic,
    MusicalSymbolCombiningSnapPizzicato,
    MusicalSymbolPedalMark,
    MusicalSymbolPedalUpMark,
    MusicalSymbolHalfPedalMark,
    MusicalSymbolGlissandoUp,
    MusicalSymbolGlissandoDown,
    MusicalSymbolWithFingernails,
    MusicalSymbolDamp,
    MusicalSymbolDampAll,
    MusicalSymbolMaxima,
    MusicalSymbolLonga,
    MusicalSymbolBrevis,
    MusicalSymbolSemibrevisWhite,
    MusicalSymbolSemibrevisBlack,
    MusicalSymbolMinima,
    MusicalSymbolMinimaBlack,
    MusicalSymbolSemiminimaWhite,
    MusicalSymbolSemiminimaBlack,
    MusicalSymbolFusaWhite,
    MusicalSymbolFusaBlack,
    MusicalSymbolLongaPerfectaRest,
    MusicalSymbolLongaImperfectaRest,
    MusicalSymbolBrevisRest,
    MusicalSymbolSemibrevisRest,
    MusicalSymbolMinimaRest,
    MusicalSymbolSemiminimaRest,
    MusicalSymbolTempusPerfectumCumProlationePerfecta,
    MusicalSymbolTempusPerfectumCumProlationeImperfecta,
    MusicalSymbolTempusPerfectumCumProlationePerfectaDiminutionDash1,
    MusicalSymbolTempusImperfectumCumProlationePerfecta,
    MusicalSymbolTempusImperfectumCumProlationeImperfecta,
    MusicalSymbolTempusImperfectumCumProlationeImperfectaDiminutionDash1,
    MusicalSymbolTempusImperfectumCumProlationeImperfectaDiminutionDash2,
    MusicalSymbolTempusImperfectumCumProlationeImperfectaDiminutionDash3,
    MusicalSymbolCroix,
    MusicalSymbolGregorianCClef,
    MusicalSymbolGregorianFClef,
    MusicalSymbolSquareB,
    MusicalSymbolVirga,
    MusicalSymbolPodatus,
    MusicalSymbolClivis,
    MusicalSymbolScandicus,
    MusicalSymbolClimacus,
    MusicalSymbolTorculus,
    MusicalSymbolPorrectus,
    MusicalSymbolPorrectusFlexus,
    MusicalSymbolScandicusFlexus,
    MusicalSymbolTorculusResupinus,
    MusicalSymbolPesSubpunctis,
    MusicalSymbolKievanCClef,
    MusicalSymbolKievanEndOfPiece,
    MusicalSymbolKievanFinalNote,
    MusicalSymbolKievanRecitativeMark,
    MusicalSymbolKievanWholeNote,
    MusicalSymbolKievanHalfNote,
    MusicalSymbolKievanQuarterNoteStemDown,
    MusicalSymbolKievanQuarterNoteStemUp,
    MusicalSymbolKievanEighthNoteStemDown,
    MusicalSymbolKievanEighthNoteStemUp,
    MusicalSymbolKievanFlatSign,
}

An enum to represent all characters in the MusicalSymbols block.

Variants
MusicalSymbolSingleBarline

\u{1d100}: '𝄀'

MusicalSymbolDoubleBarline

\u{1d101}: '𝄁'

MusicalSymbolFinalBarline

\u{1d102}: '𝄂'

MusicalSymbolReverseFinalBarline

\u{1d103}: '𝄃'

MusicalSymbolDashedBarline

\u{1d104}: '𝄄'

MusicalSymbolShortBarline

\u{1d105}: '𝄅'

MusicalSymbolLeftRepeatSign

\u{1d106}: '𝄆'

MusicalSymbolRightRepeatSign

\u{1d107}: '𝄇'

MusicalSymbolRepeatDots

\u{1d108}: '𝄈'

MusicalSymbolDalSegno

\u{1d109}: '𝄉'

MusicalSymbolDaCapo

\u{1d10a}: '𝄊'

MusicalSymbolSegno

\u{1d10b}: '𝄋'

MusicalSymbolCoda

\u{1d10c}: '𝄌'

MusicalSymbolRepeatedFigureDash1

\u{1d10d}: '𝄍'

MusicalSymbolRepeatedFigureDash2

\u{1d10e}: '𝄎'

MusicalSymbolRepeatedFigureDash3

\u{1d10f}: '𝄏'

MusicalSymbolFermata

\u{1d110}: '𝄐'

MusicalSymbolFermataBelow

\u{1d111}: '𝄑'

MusicalSymbolBreathMark

\u{1d112}: '𝄒'

MusicalSymbolCaesura

\u{1d113}: '𝄓'

MusicalSymbolBrace

\u{1d114}: '𝄔'

MusicalSymbolBracket

\u{1d115}: '𝄕'

MusicalSymbolOneDashLineStaff

\u{1d116}: '𝄖'

MusicalSymbolTwoDashLineStaff

\u{1d117}: '𝄗'

MusicalSymbolThreeDashLineStaff

\u{1d118}: '𝄘'

MusicalSymbolFourDashLineStaff

\u{1d119}: '𝄙'

MusicalSymbolFiveDashLineStaff

\u{1d11a}: '𝄚'

MusicalSymbolSixDashLineStaff

\u{1d11b}: '𝄛'

MusicalSymbolSixDashStringFretboard

\u{1d11c}: '𝄜'

MusicalSymbolFourDashStringFretboard

\u{1d11d}: '𝄝'

MusicalSymbolGClef

\u{1d11e}: '𝄞'

MusicalSymbolGClefOttavaAlta

\u{1d11f}: '𝄟'

MusicalSymbolGClefOttavaBassa

\u{1d120}: '𝄠'

MusicalSymbolCClef

\u{1d121}: '𝄡'

MusicalSymbolFClef

\u{1d122}: '𝄢'

MusicalSymbolFClefOttavaAlta

\u{1d123}: '𝄣'

MusicalSymbolFClefOttavaBassa

\u{1d124}: '𝄤'

MusicalSymbolDrumClefDash1

\u{1d125}: '𝄥'

MusicalSymbolDrumClefDash2

\u{1d126}: '𝄦'

MusicalSymbolMultipleMeasureRest

\u{1d129}: '𝄩'

MusicalSymbolDoubleSharp

\u{1d12a}: '𝄪'

MusicalSymbolDoubleFlat

\u{1d12b}: '𝄫'

MusicalSymbolFlatUp

\u{1d12c}: '𝄬'

MusicalSymbolFlatDown

\u{1d12d}: '𝄭'

MusicalSymbolNaturalUp

\u{1d12e}: '𝄮'

MusicalSymbolNaturalDown

\u{1d12f}: '𝄯'

MusicalSymbolSharpUp

\u{1d130}: '𝄰'

MusicalSymbolSharpDown

\u{1d131}: '𝄱'

MusicalSymbolQuarterToneSharp

\u{1d132}: '𝄲'

MusicalSymbolQuarterToneFlat

\u{1d133}: '𝄳'

MusicalSymbolCommonTime

\u{1d134}: '𝄴'

MusicalSymbolCutTime

\u{1d135}: '𝄵'

MusicalSymbolOttavaAlta

\u{1d136}: '𝄶'

MusicalSymbolOttavaBassa

\u{1d137}: '𝄷'

MusicalSymbolQuindicesimaAlta

\u{1d138}: '𝄸'

MusicalSymbolQuindicesimaBassa

\u{1d139}: '𝄹'

MusicalSymbolMultiRest

\u{1d13a}: '𝄺'

MusicalSymbolWholeRest

\u{1d13b}: '𝄻'

MusicalSymbolHalfRest

\u{1d13c}: '𝄼'

MusicalSymbolQuarterRest

\u{1d13d}: '𝄽'

MusicalSymbolEighthRest

\u{1d13e}: '𝄾'

MusicalSymbolSixteenthRest

\u{1d13f}: '𝄿'

MusicalSymbolThirtyDashSecondRest

\u{1d140}: '𝅀'

MusicalSymbolSixtyDashFourthRest

\u{1d141}: '𝅁'

MusicalSymbolOneHundredTwentyDashEighthRest

\u{1d142}: '𝅂'

MusicalSymbolXNotehead

\u{1d143}: '𝅃'

MusicalSymbolPlusNotehead

\u{1d144}: '𝅄'

MusicalSymbolCircleXNotehead

\u{1d145}: '𝅅'

MusicalSymbolSquareNoteheadWhite

\u{1d146}: '𝅆'

MusicalSymbolSquareNoteheadBlack

\u{1d147}: '𝅇'

MusicalSymbolTriangleNoteheadUpWhite

\u{1d148}: '𝅈'

MusicalSymbolTriangleNoteheadUpBlack

\u{1d149}: '𝅉'

MusicalSymbolTriangleNoteheadLeftWhite

\u{1d14a}: '𝅊'

MusicalSymbolTriangleNoteheadLeftBlack

\u{1d14b}: '𝅋'

MusicalSymbolTriangleNoteheadRightWhite

\u{1d14c}: '𝅌'

MusicalSymbolTriangleNoteheadRightBlack

\u{1d14d}: '𝅍'

MusicalSymbolTriangleNoteheadDownWhite

\u{1d14e}: '𝅎'

MusicalSymbolTriangleNoteheadDownBlack

\u{1d14f}: '𝅏'

MusicalSymbolTriangleNoteheadUpRightWhite

\u{1d150}: '𝅐'

MusicalSymbolTriangleNoteheadUpRightBlack

\u{1d151}: '𝅑'

MusicalSymbolMoonNoteheadWhite

\u{1d152}: '𝅒'

MusicalSymbolMoonNoteheadBlack

\u{1d153}: '𝅓'

MusicalSymbolTriangleDashRoundNoteheadDownWhite

\u{1d154}: '𝅔'

MusicalSymbolTriangleDashRoundNoteheadDownBlack

\u{1d155}: '𝅕'

MusicalSymbolParenthesisNotehead

\u{1d156}: '𝅖'

MusicalSymbolVoidNotehead

\u{1d157}: '𝅗'

MusicalSymbolNoteheadBlack

\u{1d158}: '𝅘'

MusicalSymbolNullNotehead

\u{1d159}: '𝅙'

MusicalSymbolClusterNoteheadWhite

\u{1d15a}: '𝅚'

MusicalSymbolClusterNoteheadBlack

\u{1d15b}: '𝅛'

MusicalSymbolBreve

\u{1d15c}: '𝅜'

MusicalSymbolWholeNote

\u{1d15d}: '𝅝'

MusicalSymbolHalfNote

\u{1d15e}: '𝅗𝅥'

MusicalSymbolQuarterNote

\u{1d15f}: '𝅘𝅥'

MusicalSymbolEighthNote

\u{1d160}: '𝅘𝅥𝅮'

MusicalSymbolSixteenthNote

\u{1d161}: '𝅘𝅥𝅯'

MusicalSymbolThirtyDashSecondNote

\u{1d162}: '𝅘𝅥𝅰'

MusicalSymbolSixtyDashFourthNote

\u{1d163}: '𝅘𝅥𝅱'

MusicalSymbolOneHundredTwentyDashEighthNote

\u{1d164}: '𝅘𝅥𝅲'

MusicalSymbolCombiningStem

\u{1d165}: '𝅥'

MusicalSymbolCombiningSprechgesangStem

\u{1d166}: '𝅦'

MusicalSymbolCombiningTremoloDash1

\u{1d167}: '𝅧'

MusicalSymbolCombiningTremoloDash2

\u{1d168}: '𝅨'

MusicalSymbolCombiningTremoloDash3

\u{1d169}: '𝅩'

MusicalSymbolFingeredTremoloDash1

\u{1d16a}: '𝅪'

MusicalSymbolFingeredTremoloDash2

\u{1d16b}: '𝅫'

MusicalSymbolFingeredTremoloDash3

\u{1d16c}: '𝅬'

MusicalSymbolCombiningAugmentationDot

\u{1d16d}: '𝅭'

MusicalSymbolCombiningFlagDash1

\u{1d16e}: '𝅮'

MusicalSymbolCombiningFlagDash2

\u{1d16f}: '𝅯'

MusicalSymbolCombiningFlagDash3

\u{1d170}: '𝅰'

MusicalSymbolCombiningFlagDash4

\u{1d171}: '𝅱'

MusicalSymbolCombiningFlagDash5

\u{1d172}: '𝅲'

MusicalSymbolBeginBeam

\u{1d173}: '𝅳'

MusicalSymbolEndBeam

\u{1d174}: '𝅴'

MusicalSymbolBeginTie

\u{1d175}: '𝅵'

MusicalSymbolEndTie

\u{1d176}: '𝅶'

MusicalSymbolBeginSlur

\u{1d177}: '𝅷'

MusicalSymbolEndSlur

\u{1d178}: '𝅸'

MusicalSymbolBeginPhrase

\u{1d179}: '𝅹'

MusicalSymbolEndPhrase

\u{1d17a}: '𝅺'

MusicalSymbolCombiningAccent

\u{1d17b}: '𝅻'

MusicalSymbolCombiningStaccato

\u{1d17c}: '𝅼'

MusicalSymbolCombiningTenuto

\u{1d17d}: '𝅽'

MusicalSymbolCombiningStaccatissimo

\u{1d17e}: '𝅾'

MusicalSymbolCombiningMarcato

\u{1d17f}: '𝅿'

MusicalSymbolCombiningMarcatoDashStaccato

\u{1d180}: '𝆀'

MusicalSymbolCombiningAccentDashStaccato

\u{1d181}: '𝆁'

MusicalSymbolCombiningLoure

\u{1d182}: '𝆂'

MusicalSymbolArpeggiatoUp

\u{1d183}: '𝆃'

MusicalSymbolArpeggiatoDown

\u{1d184}: '𝆄'

MusicalSymbolCombiningDoit

\u{1d185}: '𝆅'

MusicalSymbolCombiningRip

\u{1d186}: '𝆆'

MusicalSymbolCombiningFlip

\u{1d187}: '𝆇'

MusicalSymbolCombiningSmear

\u{1d188}: '𝆈'

MusicalSymbolCombiningBend

\u{1d189}: '𝆉'

MusicalSymbolCombiningDoubleTongue

\u{1d18a}: '𝆊'

MusicalSymbolCombiningTripleTongue

\u{1d18b}: '𝆋'

MusicalSymbolRinforzando

\u{1d18c}: '𝆌'

MusicalSymbolSubito

\u{1d18d}: '𝆍'

MusicalSymbolZ

\u{1d18e}: '𝆎'

MusicalSymbolPiano

\u{1d18f}: '𝆏'

MusicalSymbolMezzo

\u{1d190}: '𝆐'

MusicalSymbolForte

\u{1d191}: '𝆑'

MusicalSymbolCrescendo

\u{1d192}: '𝆒'

MusicalSymbolDecrescendo

\u{1d193}: '𝆓'

MusicalSymbolGraceNoteSlash

\u{1d194}: '𝆔'

MusicalSymbolGraceNoteNoSlash

\u{1d195}: '𝆕'

MusicalSymbolTr

\u{1d196}: '𝆖'

MusicalSymbolTurn

\u{1d197}: '𝆗'

MusicalSymbolInvertedTurn

\u{1d198}: '𝆘'

MusicalSymbolTurnSlash

\u{1d199}: '𝆙'

MusicalSymbolTurnUp

\u{1d19a}: '𝆚'

MusicalSymbolOrnamentStrokeDash1

\u{1d19b}: '𝆛'

MusicalSymbolOrnamentStrokeDash2

\u{1d19c}: '𝆜'

MusicalSymbolOrnamentStrokeDash3

\u{1d19d}: '𝆝'

MusicalSymbolOrnamentStrokeDash4

\u{1d19e}: '𝆞'

MusicalSymbolOrnamentStrokeDash5

\u{1d19f}: '𝆟'

MusicalSymbolOrnamentStrokeDash6

\u{1d1a0}: '𝆠'

MusicalSymbolOrnamentStrokeDash7

\u{1d1a1}: '𝆡'

MusicalSymbolOrnamentStrokeDash8

\u{1d1a2}: '𝆢'

MusicalSymbolOrnamentStrokeDash9

\u{1d1a3}: '𝆣'

MusicalSymbolOrnamentStrokeDash10

\u{1d1a4}: '𝆤'

MusicalSymbolOrnamentStrokeDash11

\u{1d1a5}: '𝆥'

MusicalSymbolHauptstimme

\u{1d1a6}: '𝆦'

MusicalSymbolNebenstimme

\u{1d1a7}: '𝆧'

MusicalSymbolEndOfStimme

\u{1d1a8}: '𝆨'

MusicalSymbolDegreeSlash

\u{1d1a9}: '𝆩'

MusicalSymbolCombiningDownBow

\u{1d1aa}: '𝆪'

MusicalSymbolCombiningUpBow

\u{1d1ab}: '𝆫'

MusicalSymbolCombiningHarmonic

\u{1d1ac}: '𝆬'

MusicalSymbolCombiningSnapPizzicato

\u{1d1ad}: '𝆭'

MusicalSymbolPedalMark

\u{1d1ae}: '𝆮'

MusicalSymbolPedalUpMark

\u{1d1af}: '𝆯'

MusicalSymbolHalfPedalMark

\u{1d1b0}: '𝆰'

MusicalSymbolGlissandoUp

\u{1d1b1}: '𝆱'

MusicalSymbolGlissandoDown

\u{1d1b2}: '𝆲'

MusicalSymbolWithFingernails

\u{1d1b3}: '𝆳'

MusicalSymbolDamp

\u{1d1b4}: '𝆴'

MusicalSymbolDampAll

\u{1d1b5}: '𝆵'

MusicalSymbolMaxima

\u{1d1b6}: '𝆶'

MusicalSymbolLonga

\u{1d1b7}: '𝆷'

MusicalSymbolBrevis

\u{1d1b8}: '𝆸'

MusicalSymbolSemibrevisWhite

\u{1d1b9}: '𝆹'

MusicalSymbolSemibrevisBlack

\u{1d1ba}: '𝆺'

MusicalSymbolMinima

\u{1d1bb}: '𝆹𝅥'

MusicalSymbolMinimaBlack

\u{1d1bc}: '𝆺𝅥'

MusicalSymbolSemiminimaWhite

\u{1d1bd}: '𝆹𝅥𝅮'

MusicalSymbolSemiminimaBlack

\u{1d1be}: '𝆺𝅥𝅮'

MusicalSymbolFusaWhite

\u{1d1bf}: '𝆹𝅥𝅯'

MusicalSymbolFusaBlack

\u{1d1c0}: '𝆺𝅥𝅯'

MusicalSymbolLongaPerfectaRest

\u{1d1c1}: '𝇁'

MusicalSymbolLongaImperfectaRest

\u{1d1c2}: '𝇂'

MusicalSymbolBrevisRest

\u{1d1c3}: '𝇃'

MusicalSymbolSemibrevisRest

\u{1d1c4}: '𝇄'

MusicalSymbolMinimaRest

\u{1d1c5}: '𝇅'

MusicalSymbolSemiminimaRest

\u{1d1c6}: '𝇆'

MusicalSymbolTempusPerfectumCumProlationePerfecta

\u{1d1c7}: '𝇇'

MusicalSymbolTempusPerfectumCumProlationeImperfecta

\u{1d1c8}: '𝇈'

MusicalSymbolTempusPerfectumCumProlationePerfectaDiminutionDash1

\u{1d1c9}: '𝇉'

MusicalSymbolTempusImperfectumCumProlationePerfecta

\u{1d1ca}: '𝇊'

MusicalSymbolTempusImperfectumCumProlationeImperfecta

\u{1d1cb}: '𝇋'

MusicalSymbolTempusImperfectumCumProlationeImperfectaDiminutionDash1

\u{1d1cc}: '𝇌'

MusicalSymbolTempusImperfectumCumProlationeImperfectaDiminutionDash2

\u{1d1cd}: '𝇍'

MusicalSymbolTempusImperfectumCumProlationeImperfectaDiminutionDash3

\u{1d1ce}: '𝇎'

MusicalSymbolCroix

\u{1d1cf}: '𝇏'

MusicalSymbolGregorianCClef

\u{1d1d0}: '𝇐'

MusicalSymbolGregorianFClef

\u{1d1d1}: '𝇑'

MusicalSymbolSquareB

\u{1d1d2}: '𝇒'

MusicalSymbolVirga

\u{1d1d3}: '𝇓'

MusicalSymbolPodatus

\u{1d1d4}: '𝇔'

MusicalSymbolClivis

\u{1d1d5}: '𝇕'

MusicalSymbolScandicus

\u{1d1d6}: '𝇖'

MusicalSymbolClimacus

\u{1d1d7}: '𝇗'

MusicalSymbolTorculus

\u{1d1d8}: '𝇘'

MusicalSymbolPorrectus

\u{1d1d9}: '𝇙'

MusicalSymbolPorrectusFlexus

\u{1d1da}: '𝇚'

MusicalSymbolScandicusFlexus

\u{1d1db}: '𝇛'

MusicalSymbolTorculusResupinus

\u{1d1dc}: '𝇜'

MusicalSymbolPesSubpunctis

\u{1d1dd}: '𝇝'

MusicalSymbolKievanCClef

\u{1d1de}: '𝇞'

MusicalSymbolKievanEndOfPiece

\u{1d1df}: '𝇟'

MusicalSymbolKievanFinalNote

\u{1d1e0}: '𝇠'

MusicalSymbolKievanRecitativeMark

\u{1d1e1}: '𝇡'

MusicalSymbolKievanWholeNote

\u{1d1e2}: '𝇢'

MusicalSymbolKievanHalfNote

\u{1d1e3}: '𝇣'

MusicalSymbolKievanQuarterNoteStemDown

\u{1d1e4}: '𝇤'

MusicalSymbolKievanQuarterNoteStemUp

\u{1d1e5}: '𝇥'

MusicalSymbolKievanEighthNoteStemDown

\u{1d1e6}: '𝇦'

MusicalSymbolKievanEighthNoteStemUp

\u{1d1e7}: '𝇧'

MusicalSymbolKievanFlatSign

\u{1d1e8}: '𝇨'

Methods

impl MusicalSymbols

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case