Enum unicode_types::myanmar_extended_b::MyanmarExtendedB

pub enum MyanmarExtendedB {
    MyanmarLetterShanGha,
    MyanmarLetterShanCha,
    MyanmarLetterShanJha,
    MyanmarLetterShanNna,
    MyanmarLetterShanBha,
    MyanmarSignShanSaw,
    MyanmarModifierLetterShanReduplication,
    MyanmarLetterTaiLaingNya,
    MyanmarLetterTaiLaingFa,
    MyanmarLetterTaiLaingGa,
    MyanmarLetterTaiLaingGha,
    MyanmarLetterTaiLaingJa,
    MyanmarLetterTaiLaingJha,
    MyanmarLetterTaiLaingDda,
    MyanmarLetterTaiLaingDdha,
    MyanmarLetterTaiLaingNna,
    MyanmarTaiLaingDigitZero,
    MyanmarTaiLaingDigitOne,
    MyanmarTaiLaingDigitTwo,
    MyanmarTaiLaingDigitThree,
    MyanmarTaiLaingDigitFour,
    MyanmarTaiLaingDigitFive,
    MyanmarTaiLaingDigitSix,
    MyanmarTaiLaingDigitSeven,
    MyanmarTaiLaingDigitEight,
    MyanmarTaiLaingDigitNine,
    MyanmarLetterTaiLaingLla,
    MyanmarLetterTaiLaingDa,
    MyanmarLetterTaiLaingDha,
    MyanmarLetterTaiLaingBa,
    MyanmarLetterTaiLaingBha,
}

An enum to represent all characters in the MyanmarExtendedB block.

Variants
MyanmarLetterShanGha

\u{a9e0}: 'ꧠ'

MyanmarLetterShanCha

\u{a9e1}: 'ꧡ'

MyanmarLetterShanJha

\u{a9e2}: 'ꧢ'

MyanmarLetterShanNna

\u{a9e3}: 'ꧣ'

MyanmarLetterShanBha

\u{a9e4}: 'ꧤ'

MyanmarSignShanSaw

\u{a9e5}: 'ꧥ'

MyanmarModifierLetterShanReduplication

\u{a9e6}: 'ꧦ'

MyanmarLetterTaiLaingNya

\u{a9e7}: 'ꧧ'

MyanmarLetterTaiLaingFa

\u{a9e8}: 'ꧨ'

MyanmarLetterTaiLaingGa

\u{a9e9}: 'ꧩ'

MyanmarLetterTaiLaingGha

\u{a9ea}: 'ꧪ'

MyanmarLetterTaiLaingJa

\u{a9eb}: 'ꧫ'

MyanmarLetterTaiLaingJha

\u{a9ec}: 'ꧬ'

MyanmarLetterTaiLaingDda

\u{a9ed}: 'ꧭ'

MyanmarLetterTaiLaingDdha

\u{a9ee}: 'ꧮ'

MyanmarLetterTaiLaingNna

\u{a9ef}: 'ꧯ'

MyanmarTaiLaingDigitZero

\u{a9f0}: '꧰'

MyanmarTaiLaingDigitOne

\u{a9f1}: '꧱'

MyanmarTaiLaingDigitTwo

\u{a9f2}: '꧲'

MyanmarTaiLaingDigitThree

\u{a9f3}: '꧳'

MyanmarTaiLaingDigitFour

\u{a9f4}: '꧴'

MyanmarTaiLaingDigitFive

\u{a9f5}: '꧵'

MyanmarTaiLaingDigitSix

\u{a9f6}: '꧶'

MyanmarTaiLaingDigitSeven

\u{a9f7}: '꧷'

MyanmarTaiLaingDigitEight

\u{a9f8}: '꧸'

MyanmarTaiLaingDigitNine

\u{a9f9}: '꧹'

MyanmarLetterTaiLaingLla

\u{a9fa}: 'ꧺ'

MyanmarLetterTaiLaingDa

\u{a9fb}: 'ꧻ'

MyanmarLetterTaiLaingDha

\u{a9fc}: 'ꧼ'

MyanmarLetterTaiLaingBa

\u{a9fd}: 'ꧽ'

MyanmarLetterTaiLaingBha

\u{a9fe}: 'ꧾ'

Methods

impl MyanmarExtendedB

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case