Enum unicode_types::nushu::Nushu

pub enum Nushu {
    CharacterDash1b170,
    CharacterDash1b171,
    CharacterDash1b172,
    CharacterDash1b173,
    CharacterDash1b174,
    CharacterDash1b175,
    CharacterDash1b176,
    CharacterDash1b177,
    CharacterDash1b178,
    CharacterDash1b179,
    CharacterDash1b17a,
    CharacterDash1b17b,
    CharacterDash1b17c,
    CharacterDash1b17d,
    CharacterDash1b17e,
    CharacterDash1b17f,
    CharacterDash1b180,
    CharacterDash1b181,
    CharacterDash1b182,
    CharacterDash1b183,
    CharacterDash1b184,
    CharacterDash1b185,
    CharacterDash1b186,
    CharacterDash1b187,
    CharacterDash1b188,
    CharacterDash1b189,
    CharacterDash1b18a,
    CharacterDash1b18b,
    CharacterDash1b18c,
    CharacterDash1b18d,
    CharacterDash1b18e,
    CharacterDash1b18f,
    CharacterDash1b190,
    CharacterDash1b191,
    CharacterDash1b192,
    CharacterDash1b193,
    CharacterDash1b194,
    CharacterDash1b195,
    CharacterDash1b196,
    CharacterDash1b197,
    CharacterDash1b198,
    CharacterDash1b199,
    CharacterDash1b19a,
    CharacterDash1b19b,
    CharacterDash1b19c,
    CharacterDash1b19d,
    CharacterDash1b19e,
    CharacterDash1b19f,
    CharacterDash1b1a0,
    CharacterDash1b1a1,
    CharacterDash1b1a2,
    CharacterDash1b1a3,
    CharacterDash1b1a4,
    CharacterDash1b1a5,
    CharacterDash1b1a6,
    CharacterDash1b1a7,
    CharacterDash1b1a8,
    CharacterDash1b1a9,
    CharacterDash1b1aa,
    CharacterDash1b1ab,
    CharacterDash1b1ac,
    CharacterDash1b1ad,
    CharacterDash1b1ae,
    CharacterDash1b1af,
    CharacterDash1b1b0,
    CharacterDash1b1b1,
    CharacterDash1b1b2,
    CharacterDash1b1b3,
    CharacterDash1b1b4,
    CharacterDash1b1b5,
    CharacterDash1b1b6,
    CharacterDash1b1b7,
    CharacterDash1b1b8,
    CharacterDash1b1b9,
    CharacterDash1b1ba,
    CharacterDash1b1bb,
    CharacterDash1b1bc,
    CharacterDash1b1bd,
    CharacterDash1b1be,
    CharacterDash1b1bf,
    CharacterDash1b1c0,
    CharacterDash1b1c1,
    CharacterDash1b1c2,
    CharacterDash1b1c3,
    CharacterDash1b1c4,
    CharacterDash1b1c5,
    CharacterDash1b1c6,
    CharacterDash1b1c7,
    CharacterDash1b1c8,
    CharacterDash1b1c9,
    CharacterDash1b1ca,
    CharacterDash1b1cb,
    CharacterDash1b1cc,
    CharacterDash1b1cd,
    CharacterDash1b1ce,
    CharacterDash1b1cf,
    CharacterDash1b1d0,
    CharacterDash1b1d1,
    CharacterDash1b1d2,
    CharacterDash1b1d3,
    CharacterDash1b1d4,
    CharacterDash1b1d5,
    CharacterDash1b1d6,
    CharacterDash1b1d7,
    CharacterDash1b1d8,
    CharacterDash1b1d9,
    CharacterDash1b1da,
    CharacterDash1b1db,
    CharacterDash1b1dc,
    CharacterDash1b1dd,
    CharacterDash1b1de,
    CharacterDash1b1df,
    CharacterDash1b1e0,
    CharacterDash1b1e1,
    CharacterDash1b1e2,
    CharacterDash1b1e3,
    CharacterDash1b1e4,
    CharacterDash1b1e5,
    CharacterDash1b1e6,
    CharacterDash1b1e7,
    CharacterDash1b1e8,
    CharacterDash1b1e9,
    CharacterDash1b1ea,
    CharacterDash1b1eb,
    CharacterDash1b1ec,
    CharacterDash1b1ed,
    CharacterDash1b1ee,
    CharacterDash1b1ef,
    CharacterDash1b1f0,
    CharacterDash1b1f1,
    CharacterDash1b1f2,
    CharacterDash1b1f3,
    CharacterDash1b1f4,
    CharacterDash1b1f5,
    CharacterDash1b1f6,
    CharacterDash1b1f7,
    CharacterDash1b1f8,
    CharacterDash1b1f9,
    CharacterDash1b1fa,
    CharacterDash1b1fb,
    CharacterDash1b1fc,
    CharacterDash1b1fd,
    CharacterDash1b1fe,
    CharacterDash1b1ff,
    CharacterDash1b200,
    CharacterDash1b201,
    CharacterDash1b202,
    CharacterDash1b203,
    CharacterDash1b204,
    CharacterDash1b205,
    CharacterDash1b206,
    CharacterDash1b207,
    CharacterDash1b208,
    CharacterDash1b209,
    CharacterDash1b20a,
    CharacterDash1b20b,
    CharacterDash1b20c,
    CharacterDash1b20d,
    CharacterDash1b20e,
    CharacterDash1b20f,
    CharacterDash1b210,
    CharacterDash1b211,
    CharacterDash1b212,
    CharacterDash1b213,
    CharacterDash1b214,
    CharacterDash1b215,
    CharacterDash1b216,
    CharacterDash1b217,
    CharacterDash1b218,
    CharacterDash1b219,
    CharacterDash1b21a,
    CharacterDash1b21b,
    CharacterDash1b21c,
    CharacterDash1b21d,
    CharacterDash1b21e,
    CharacterDash1b21f,
    CharacterDash1b220,
    CharacterDash1b221,
    CharacterDash1b222,
    CharacterDash1b223,
    CharacterDash1b224,
    CharacterDash1b225,
    CharacterDash1b226,
    CharacterDash1b227,
    CharacterDash1b228,
    CharacterDash1b229,
    CharacterDash1b22a,
    CharacterDash1b22b,
    CharacterDash1b22c,
    CharacterDash1b22d,
    CharacterDash1b22e,
    CharacterDash1b22f,
    CharacterDash1b230,
    CharacterDash1b231,
    CharacterDash1b232,
    CharacterDash1b233,
    CharacterDash1b234,
    CharacterDash1b235,
    CharacterDash1b236,
    CharacterDash1b237,
    CharacterDash1b238,
    CharacterDash1b239,
    CharacterDash1b23a,
    CharacterDash1b23b,
    CharacterDash1b23c,
    CharacterDash1b23d,
    CharacterDash1b23e,
    CharacterDash1b23f,
    CharacterDash1b240,
    CharacterDash1b241,
    CharacterDash1b242,
    CharacterDash1b243,
    CharacterDash1b244,
    CharacterDash1b245,
    CharacterDash1b246,
    CharacterDash1b247,
    CharacterDash1b248,
    CharacterDash1b249,
    CharacterDash1b24a,
    CharacterDash1b24b,
    CharacterDash1b24c,
    CharacterDash1b24d,
    CharacterDash1b24e,
    CharacterDash1b24f,
    CharacterDash1b250,
    CharacterDash1b251,
    CharacterDash1b252,
    CharacterDash1b253,
    CharacterDash1b254,
    CharacterDash1b255,
    CharacterDash1b256,
    CharacterDash1b257,
    CharacterDash1b258,
    CharacterDash1b259,
    CharacterDash1b25a,
    CharacterDash1b25b,
    CharacterDash1b25c,
    CharacterDash1b25d,
    CharacterDash1b25e,
    CharacterDash1b25f,
    CharacterDash1b260,
    CharacterDash1b261,
    CharacterDash1b262,
    CharacterDash1b263,
    CharacterDash1b264,
    CharacterDash1b265,
    CharacterDash1b266,
    CharacterDash1b267,
    CharacterDash1b268,
    CharacterDash1b269,
    CharacterDash1b26a,
    CharacterDash1b26b,
    CharacterDash1b26c,
    CharacterDash1b26d,
    CharacterDash1b26e,
    CharacterDash1b26f,
    CharacterDash1b270,
    CharacterDash1b271,
    CharacterDash1b272,
    CharacterDash1b273,
    CharacterDash1b274,
    CharacterDash1b275,
    CharacterDash1b276,
    CharacterDash1b277,
    CharacterDash1b278,
    CharacterDash1b279,
    CharacterDash1b27a,
    CharacterDash1b27b,
    CharacterDash1b27c,
    CharacterDash1b27d,
    CharacterDash1b27e,
    CharacterDash1b27f,
    CharacterDash1b280,
    CharacterDash1b281,
    CharacterDash1b282,
    CharacterDash1b283,
    CharacterDash1b284,
    CharacterDash1b285,
    CharacterDash1b286,
    CharacterDash1b287,
    CharacterDash1b288,
    CharacterDash1b289,
    CharacterDash1b28a,
    CharacterDash1b28b,
    CharacterDash1b28c,
    CharacterDash1b28d,
    CharacterDash1b28e,
    CharacterDash1b28f,
    CharacterDash1b290,
    CharacterDash1b291,
    CharacterDash1b292,
    CharacterDash1b293,
    CharacterDash1b294,
    CharacterDash1b295,
    CharacterDash1b296,
    CharacterDash1b297,
    CharacterDash1b298,
    CharacterDash1b299,
    CharacterDash1b29a,
    CharacterDash1b29b,
    CharacterDash1b29c,
    CharacterDash1b29d,
    CharacterDash1b29e,
    CharacterDash1b29f,
    CharacterDash1b2a0,
    CharacterDash1b2a1,
    CharacterDash1b2a2,
    CharacterDash1b2a3,
    CharacterDash1b2a4,
    CharacterDash1b2a5,
    CharacterDash1b2a6,
    CharacterDash1b2a7,
    CharacterDash1b2a8,
    CharacterDash1b2a9,
    CharacterDash1b2aa,
    CharacterDash1b2ab,
    CharacterDash1b2ac,
    CharacterDash1b2ad,
    CharacterDash1b2ae,
    CharacterDash1b2af,
    CharacterDash1b2b0,
    CharacterDash1b2b1,
    CharacterDash1b2b2,
    CharacterDash1b2b3,
    CharacterDash1b2b4,
    CharacterDash1b2b5,
    CharacterDash1b2b6,
    CharacterDash1b2b7,
    CharacterDash1b2b8,
    CharacterDash1b2b9,
    CharacterDash1b2ba,
    CharacterDash1b2bb,
    CharacterDash1b2bc,
    CharacterDash1b2bd,
    CharacterDash1b2be,
    CharacterDash1b2bf,
    CharacterDash1b2c0,
    CharacterDash1b2c1,
    CharacterDash1b2c2,
    CharacterDash1b2c3,
    CharacterDash1b2c4,
    CharacterDash1b2c5,
    CharacterDash1b2c6,
    CharacterDash1b2c7,
    CharacterDash1b2c8,
    CharacterDash1b2c9,
    CharacterDash1b2ca,
    CharacterDash1b2cb,
    CharacterDash1b2cc,
    CharacterDash1b2cd,
    CharacterDash1b2ce,
    CharacterDash1b2cf,
    CharacterDash1b2d0,
    CharacterDash1b2d1,
    CharacterDash1b2d2,
    CharacterDash1b2d3,
    CharacterDash1b2d4,
    CharacterDash1b2d5,
    CharacterDash1b2d6,
    CharacterDash1b2d7,
    CharacterDash1b2d8,
    CharacterDash1b2d9,
    CharacterDash1b2da,
    CharacterDash1b2db,
    CharacterDash1b2dc,
    CharacterDash1b2dd,
    CharacterDash1b2de,
    CharacterDash1b2df,
    CharacterDash1b2e0,
    CharacterDash1b2e1,
    CharacterDash1b2e2,
    CharacterDash1b2e3,
    CharacterDash1b2e4,
    CharacterDash1b2e5,
    CharacterDash1b2e6,
    CharacterDash1b2e7,
    CharacterDash1b2e8,
    CharacterDash1b2e9,
    CharacterDash1b2ea,
    CharacterDash1b2eb,
    CharacterDash1b2ec,
    CharacterDash1b2ed,
    CharacterDash1b2ee,
    CharacterDash1b2ef,
    CharacterDash1b2f0,
    CharacterDash1b2f1,
    CharacterDash1b2f2,
    CharacterDash1b2f3,
    CharacterDash1b2f4,
    CharacterDash1b2f5,
    CharacterDash1b2f6,
    CharacterDash1b2f7,
    CharacterDash1b2f8,
    CharacterDash1b2f9,
    CharacterDash1b2fa,
    CharacterDash1b2fb,
}

An enum to represent all characters in the Nushu block.

Variants
CharacterDash1b170

\u{1b170}: 'ğ›…°'

CharacterDash1b171

\u{1b171}: 'ğ›…±'

CharacterDash1b172

\u{1b172}: 'ğ›…²'

CharacterDash1b173

\u{1b173}: 'ğ›…³'

CharacterDash1b174

\u{1b174}: 'ğ›…´'

CharacterDash1b175

\u{1b175}: 'ğ›…µ'

CharacterDash1b176

\u{1b176}: 'ğ›…¶'

CharacterDash1b177

\u{1b177}: 'ğ›…·'

CharacterDash1b178

\u{1b178}: 'ğ›…¸'

CharacterDash1b179

\u{1b179}: 'ğ›…¹'

CharacterDash1b17a

\u{1b17a}: 'ğ›…º'

CharacterDash1b17b

\u{1b17b}: 'ğ›…»'

CharacterDash1b17c

\u{1b17c}: 'ğ›…¼'

CharacterDash1b17d

\u{1b17d}: 'ğ›…½'

CharacterDash1b17e

\u{1b17e}: 'ğ›…¾'

CharacterDash1b17f

\u{1b17f}: 'ğ›…¿'

CharacterDash1b180

\u{1b180}: 'ğ›†€'

CharacterDash1b181

\u{1b181}: 'ğ›†'

CharacterDash1b182

\u{1b182}: 'ğ›†‚'

CharacterDash1b183

\u{1b183}: 'ğ›†ƒ'

CharacterDash1b184

\u{1b184}: 'ğ›†„'

CharacterDash1b185

\u{1b185}: 'ğ›†…'

CharacterDash1b186

\u{1b186}: 'ğ›††'

CharacterDash1b187

\u{1b187}: 'ğ›†‡'

CharacterDash1b188

\u{1b188}: 'ğ›†ˆ'

CharacterDash1b189

\u{1b189}: 'ğ›†‰'

CharacterDash1b18a

\u{1b18a}: 'ğ›†Š'

CharacterDash1b18b

\u{1b18b}: 'ğ›†‹'

CharacterDash1b18c

\u{1b18c}: 'ğ›†Œ'

CharacterDash1b18d

\u{1b18d}: 'ğ›†'

CharacterDash1b18e

\u{1b18e}: 'ğ›†'

CharacterDash1b18f

\u{1b18f}: 'ğ›†'

CharacterDash1b190

\u{1b190}: 'ğ›†'

CharacterDash1b191

\u{1b191}: 'ğ›†‘'

CharacterDash1b192

\u{1b192}: 'ğ›†’'

CharacterDash1b193

\u{1b193}: 'ğ›†“'

CharacterDash1b194

\u{1b194}: 'ğ›†”'

CharacterDash1b195

\u{1b195}: 'ğ›†•'

CharacterDash1b196

\u{1b196}: 'ğ›†–'

CharacterDash1b197

\u{1b197}: 'ğ›†—'

CharacterDash1b198

\u{1b198}: 'ğ›†˜'

CharacterDash1b199

\u{1b199}: 'ğ›†™'

CharacterDash1b19a

\u{1b19a}: 'ğ›†š'

CharacterDash1b19b

\u{1b19b}: 'ğ›†›'

CharacterDash1b19c

\u{1b19c}: 'ğ›†œ'

CharacterDash1b19d

\u{1b19d}: 'ğ›†'

CharacterDash1b19e

\u{1b19e}: 'ğ›†'

CharacterDash1b19f

\u{1b19f}: 'ğ›†Ÿ'

CharacterDash1b1a0

\u{1b1a0}: 'ğ›† '

CharacterDash1b1a1

\u{1b1a1}: 'ğ›†¡'

CharacterDash1b1a2

\u{1b1a2}: 'ğ›†¢'

CharacterDash1b1a3

\u{1b1a3}: 'ğ›†£'

CharacterDash1b1a4

\u{1b1a4}: 'ğ›†¤'

CharacterDash1b1a5

\u{1b1a5}: 'ğ›†¥'

CharacterDash1b1a6

\u{1b1a6}: 'ğ›†¦'

CharacterDash1b1a7

\u{1b1a7}: 'ğ›†'

CharacterDash1b1a8

\u{1b1a8}: 'ğ›†¨'

CharacterDash1b1a9

\u{1b1a9}: 'ğ›†©'

CharacterDash1b1aa

\u{1b1aa}: 'ğ›†ª'

CharacterDash1b1ab

\u{1b1ab}: 'ğ›†«'

CharacterDash1b1ac

\u{1b1ac}: 'ğ›†¬'

CharacterDash1b1ad

\u{1b1ad}: 'ğ›†­'

CharacterDash1b1ae

\u{1b1ae}: 'ğ›†®'

CharacterDash1b1af

\u{1b1af}: 'ğ›†¯'

CharacterDash1b1b0

\u{1b1b0}: 'ğ›†°'

CharacterDash1b1b1

\u{1b1b1}: 'ğ›†±'

CharacterDash1b1b2

\u{1b1b2}: 'ğ›†²'

CharacterDash1b1b3

\u{1b1b3}: 'ğ›†³'

CharacterDash1b1b4

\u{1b1b4}: 'ğ›†´'

CharacterDash1b1b5

\u{1b1b5}: 'ğ›†µ'

CharacterDash1b1b6

\u{1b1b6}: 'ğ›†¶'

CharacterDash1b1b7

\u{1b1b7}: 'ğ›†·'

CharacterDash1b1b8

\u{1b1b8}: 'ğ›†¸'

CharacterDash1b1b9

\u{1b1b9}: 'ğ›†¹'

CharacterDash1b1ba

\u{1b1ba}: 'ğ›†º'

CharacterDash1b1bb

\u{1b1bb}: 'ğ›†»'

CharacterDash1b1bc

\u{1b1bc}: 'ğ›†¼'

CharacterDash1b1bd

\u{1b1bd}: 'ğ›†½'

CharacterDash1b1be

\u{1b1be}: 'ğ›†¾'

CharacterDash1b1bf

\u{1b1bf}: 'ğ›†¿'

CharacterDash1b1c0

\u{1b1c0}: 'ğ›‡€'

CharacterDash1b1c1

\u{1b1c1}: 'ğ›‡'

CharacterDash1b1c2

\u{1b1c2}: 'ğ›‡‚'

CharacterDash1b1c3

\u{1b1c3}: 'ğ›‡ƒ'

CharacterDash1b1c4

\u{1b1c4}: 'ğ›‡„'

CharacterDash1b1c5

\u{1b1c5}: 'ğ›‡…'

CharacterDash1b1c6

\u{1b1c6}: 'ğ›‡†'

CharacterDash1b1c7

\u{1b1c7}: 'ğ›‡‡'

CharacterDash1b1c8

\u{1b1c8}: 'ğ›‡ˆ'

CharacterDash1b1c9

\u{1b1c9}: 'ğ›‡‰'

CharacterDash1b1ca

\u{1b1ca}: 'ğ›‡Š'

CharacterDash1b1cb

\u{1b1cb}: 'ğ›‡‹'

CharacterDash1b1cc

\u{1b1cc}: 'ğ›‡Œ'

CharacterDash1b1cd

\u{1b1cd}: 'ğ›‡'

CharacterDash1b1ce

\u{1b1ce}: 'ğ›‡'

CharacterDash1b1cf

\u{1b1cf}: 'ğ›‡'

CharacterDash1b1d0

\u{1b1d0}: 'ğ›‡'

CharacterDash1b1d1

\u{1b1d1}: 'ğ›‡‘'

CharacterDash1b1d2

\u{1b1d2}: 'ğ›‡’'

CharacterDash1b1d3

\u{1b1d3}: 'ğ›‡“'

CharacterDash1b1d4

\u{1b1d4}: 'ğ›‡”'

CharacterDash1b1d5

\u{1b1d5}: 'ğ›‡•'

CharacterDash1b1d6

\u{1b1d6}: 'ğ›‡–'

CharacterDash1b1d7

\u{1b1d7}: 'ğ›‡—'

CharacterDash1b1d8

\u{1b1d8}: 'ğ›‡˜'

CharacterDash1b1d9

\u{1b1d9}: 'ğ›‡™'

CharacterDash1b1da

\u{1b1da}: 'ğ›‡š'

CharacterDash1b1db

\u{1b1db}: 'ğ›‡›'

CharacterDash1b1dc

\u{1b1dc}: 'ğ›‡œ'

CharacterDash1b1dd

\u{1b1dd}: 'ğ›‡'

CharacterDash1b1de

\u{1b1de}: 'ğ›‡'

CharacterDash1b1df

\u{1b1df}: 'ğ›‡Ÿ'

CharacterDash1b1e0

\u{1b1e0}: 'ğ›‡ '

CharacterDash1b1e1

\u{1b1e1}: 'ğ›‡¡'

CharacterDash1b1e2

\u{1b1e2}: 'ğ›‡¢'

CharacterDash1b1e3

\u{1b1e3}: 'ğ›‡£'

CharacterDash1b1e4

\u{1b1e4}: 'ğ›‡¤'

CharacterDash1b1e5

\u{1b1e5}: 'ğ›‡¥'

CharacterDash1b1e6

\u{1b1e6}: 'ğ›‡¦'

CharacterDash1b1e7

\u{1b1e7}: 'ğ›‡'

CharacterDash1b1e8

\u{1b1e8}: 'ğ›‡¨'

CharacterDash1b1e9

\u{1b1e9}: 'ğ›‡©'

CharacterDash1b1ea

\u{1b1ea}: 'ğ›‡ª'

CharacterDash1b1eb

\u{1b1eb}: 'ğ›‡«'

CharacterDash1b1ec

\u{1b1ec}: 'ğ›‡¬'

CharacterDash1b1ed

\u{1b1ed}: 'ğ›‡­'

CharacterDash1b1ee

\u{1b1ee}: 'ğ›‡®'

CharacterDash1b1ef

\u{1b1ef}: 'ğ›‡¯'

CharacterDash1b1f0

\u{1b1f0}: 'ğ›‡°'

CharacterDash1b1f1

\u{1b1f1}: 'ğ›‡±'

CharacterDash1b1f2

\u{1b1f2}: 'ğ›‡²'

CharacterDash1b1f3

\u{1b1f3}: 'ğ›‡³'

CharacterDash1b1f4

\u{1b1f4}: 'ğ›‡´'

CharacterDash1b1f5

\u{1b1f5}: 'ğ›‡µ'

CharacterDash1b1f6

\u{1b1f6}: 'ğ›‡¶'

CharacterDash1b1f7

\u{1b1f7}: 'ğ›‡·'

CharacterDash1b1f8

\u{1b1f8}: 'ğ›‡¸'

CharacterDash1b1f9

\u{1b1f9}: 'ğ›‡¹'

CharacterDash1b1fa

\u{1b1fa}: 'ğ›‡º'

CharacterDash1b1fb

\u{1b1fb}: 'ğ›‡»'

CharacterDash1b1fc

\u{1b1fc}: 'ğ›‡¼'

CharacterDash1b1fd

\u{1b1fd}: 'ğ›‡½'

CharacterDash1b1fe

\u{1b1fe}: 'ğ›‡¾'

CharacterDash1b1ff

\u{1b1ff}: 'ğ›‡¿'

CharacterDash1b200

\u{1b200}: 'ğ›ˆ€'

CharacterDash1b201

\u{1b201}: 'ğ›ˆ'

CharacterDash1b202

\u{1b202}: 'ğ›ˆ‚'

CharacterDash1b203

\u{1b203}: 'ğ›ˆƒ'

CharacterDash1b204

\u{1b204}: 'ğ›ˆ„'

CharacterDash1b205

\u{1b205}: 'ğ›ˆ…'

CharacterDash1b206

\u{1b206}: 'ğ›ˆ†'

CharacterDash1b207

\u{1b207}: 'ğ›ˆ‡'

CharacterDash1b208

\u{1b208}: 'ğ›ˆˆ'

CharacterDash1b209

\u{1b209}: 'ğ›ˆ‰'

CharacterDash1b20a

\u{1b20a}: 'ğ›ˆŠ'

CharacterDash1b20b

\u{1b20b}: 'ğ›ˆ‹'

CharacterDash1b20c

\u{1b20c}: 'ğ›ˆŒ'

CharacterDash1b20d

\u{1b20d}: 'ğ›ˆ'

CharacterDash1b20e

\u{1b20e}: 'ğ›ˆ'

CharacterDash1b20f

\u{1b20f}: 'ğ›ˆ'

CharacterDash1b210

\u{1b210}: 'ğ›ˆ'

CharacterDash1b211

\u{1b211}: 'ğ›ˆ‘'

CharacterDash1b212

\u{1b212}: 'ğ›ˆ’'

CharacterDash1b213

\u{1b213}: 'ğ›ˆ“'

CharacterDash1b214

\u{1b214}: 'ğ›ˆ”'

CharacterDash1b215

\u{1b215}: 'ğ›ˆ•'

CharacterDash1b216

\u{1b216}: 'ğ›ˆ–'

CharacterDash1b217

\u{1b217}: 'ğ›ˆ—'

CharacterDash1b218

\u{1b218}: 'ğ›ˆ˜'

CharacterDash1b219

\u{1b219}: 'ğ›ˆ™'

CharacterDash1b21a

\u{1b21a}: 'ğ›ˆš'

CharacterDash1b21b

\u{1b21b}: 'ğ›ˆ›'

CharacterDash1b21c

\u{1b21c}: 'ğ›ˆœ'

CharacterDash1b21d

\u{1b21d}: 'ğ›ˆ'

CharacterDash1b21e

\u{1b21e}: 'ğ›ˆ'

CharacterDash1b21f

\u{1b21f}: 'ğ›ˆŸ'

CharacterDash1b220

\u{1b220}: 'ğ›ˆ '

CharacterDash1b221

\u{1b221}: 'ğ›ˆ¡'

CharacterDash1b222

\u{1b222}: 'ğ›ˆ¢'

CharacterDash1b223

\u{1b223}: 'ğ›ˆ£'

CharacterDash1b224

\u{1b224}: 'ğ›ˆ¤'

CharacterDash1b225

\u{1b225}: 'ğ›ˆ¥'

CharacterDash1b226

\u{1b226}: 'ğ›ˆ¦'

CharacterDash1b227

\u{1b227}: 'ğ›ˆ'

CharacterDash1b228

\u{1b228}: 'ğ›ˆ¨'

CharacterDash1b229

\u{1b229}: 'ğ›ˆ©'

CharacterDash1b22a

\u{1b22a}: 'ğ›ˆª'

CharacterDash1b22b

\u{1b22b}: 'ğ›ˆ«'

CharacterDash1b22c

\u{1b22c}: 'ğ›ˆ¬'

CharacterDash1b22d

\u{1b22d}: 'ğ›ˆ­'

CharacterDash1b22e

\u{1b22e}: 'ğ›ˆ®'

CharacterDash1b22f

\u{1b22f}: 'ğ›ˆ¯'

CharacterDash1b230

\u{1b230}: 'ğ›ˆ°'

CharacterDash1b231

\u{1b231}: 'ğ›ˆ±'

CharacterDash1b232

\u{1b232}: 'ğ›ˆ²'

CharacterDash1b233

\u{1b233}: 'ğ›ˆ³'

CharacterDash1b234

\u{1b234}: 'ğ›ˆ´'

CharacterDash1b235

\u{1b235}: 'ğ›ˆµ'

CharacterDash1b236

\u{1b236}: 'ğ›ˆ¶'

CharacterDash1b237

\u{1b237}: 'ğ›ˆ·'

CharacterDash1b238

\u{1b238}: 'ğ›ˆ¸'

CharacterDash1b239

\u{1b239}: 'ğ›ˆ¹'

CharacterDash1b23a

\u{1b23a}: 'ğ›ˆº'

CharacterDash1b23b

\u{1b23b}: 'ğ›ˆ»'

CharacterDash1b23c

\u{1b23c}: 'ğ›ˆ¼'

CharacterDash1b23d

\u{1b23d}: 'ğ›ˆ½'

CharacterDash1b23e

\u{1b23e}: 'ğ›ˆ¾'

CharacterDash1b23f

\u{1b23f}: 'ğ›ˆ¿'

CharacterDash1b240

\u{1b240}: 'ğ›‰€'

CharacterDash1b241

\u{1b241}: 'ğ›‰'

CharacterDash1b242

\u{1b242}: 'ğ›‰‚'

CharacterDash1b243

\u{1b243}: 'ğ›‰ƒ'

CharacterDash1b244

\u{1b244}: 'ğ›‰„'

CharacterDash1b245

\u{1b245}: 'ğ›‰…'

CharacterDash1b246

\u{1b246}: 'ğ›‰†'

CharacterDash1b247

\u{1b247}: 'ğ›‰‡'

CharacterDash1b248

\u{1b248}: 'ğ›‰ˆ'

CharacterDash1b249

\u{1b249}: 'ğ›‰‰'

CharacterDash1b24a

\u{1b24a}: 'ğ›‰Š'

CharacterDash1b24b

\u{1b24b}: 'ğ›‰‹'

CharacterDash1b24c

\u{1b24c}: 'ğ›‰Œ'

CharacterDash1b24d

\u{1b24d}: 'ğ›‰'

CharacterDash1b24e

\u{1b24e}: 'ğ›‰'

CharacterDash1b24f

\u{1b24f}: 'ğ›‰'

CharacterDash1b250

\u{1b250}: 'ğ›‰'

CharacterDash1b251

\u{1b251}: 'ğ›‰‘'

CharacterDash1b252

\u{1b252}: 'ğ›‰’'

CharacterDash1b253

\u{1b253}: 'ğ›‰“'

CharacterDash1b254

\u{1b254}: 'ğ›‰”'

CharacterDash1b255

\u{1b255}: 'ğ›‰•'

CharacterDash1b256

\u{1b256}: 'ğ›‰–'

CharacterDash1b257

\u{1b257}: 'ğ›‰—'

CharacterDash1b258

\u{1b258}: 'ğ›‰˜'

CharacterDash1b259

\u{1b259}: 'ğ›‰™'

CharacterDash1b25a

\u{1b25a}: 'ğ›‰š'

CharacterDash1b25b

\u{1b25b}: 'ğ›‰›'

CharacterDash1b25c

\u{1b25c}: 'ğ›‰œ'

CharacterDash1b25d

\u{1b25d}: 'ğ›‰'

CharacterDash1b25e

\u{1b25e}: 'ğ›‰'

CharacterDash1b25f

\u{1b25f}: 'ğ›‰Ÿ'

CharacterDash1b260

\u{1b260}: 'ğ›‰ '

CharacterDash1b261

\u{1b261}: 'ğ›‰¡'

CharacterDash1b262

\u{1b262}: 'ğ›‰¢'

CharacterDash1b263

\u{1b263}: 'ğ›‰£'

CharacterDash1b264

\u{1b264}: 'ğ›‰¤'

CharacterDash1b265

\u{1b265}: 'ğ›‰¥'

CharacterDash1b266

\u{1b266}: 'ğ›‰¦'

CharacterDash1b267

\u{1b267}: 'ğ›‰'

CharacterDash1b268

\u{1b268}: 'ğ›‰¨'

CharacterDash1b269

\u{1b269}: 'ğ›‰©'

CharacterDash1b26a

\u{1b26a}: 'ğ›‰ª'

CharacterDash1b26b

\u{1b26b}: 'ğ›‰«'

CharacterDash1b26c

\u{1b26c}: 'ğ›‰¬'

CharacterDash1b26d

\u{1b26d}: 'ğ›‰­'

CharacterDash1b26e

\u{1b26e}: 'ğ›‰®'

CharacterDash1b26f

\u{1b26f}: 'ğ›‰¯'

CharacterDash1b270

\u{1b270}: 'ğ›‰°'

CharacterDash1b271

\u{1b271}: 'ğ›‰±'

CharacterDash1b272

\u{1b272}: 'ğ›‰²'

CharacterDash1b273

\u{1b273}: 'ğ›‰³'

CharacterDash1b274

\u{1b274}: 'ğ›‰´'

CharacterDash1b275

\u{1b275}: 'ğ›‰µ'

CharacterDash1b276

\u{1b276}: 'ğ›‰¶'

CharacterDash1b277

\u{1b277}: 'ğ›‰·'

CharacterDash1b278

\u{1b278}: 'ğ›‰¸'

CharacterDash1b279

\u{1b279}: 'ğ›‰¹'

CharacterDash1b27a

\u{1b27a}: 'ğ›‰º'

CharacterDash1b27b

\u{1b27b}: 'ğ›‰»'

CharacterDash1b27c

\u{1b27c}: 'ğ›‰¼'

CharacterDash1b27d

\u{1b27d}: 'ğ›‰½'

CharacterDash1b27e

\u{1b27e}: 'ğ›‰¾'

CharacterDash1b27f

\u{1b27f}: 'ğ›‰¿'

CharacterDash1b280

\u{1b280}: 'ğ›Š€'

CharacterDash1b281

\u{1b281}: 'ğ›Š'

CharacterDash1b282

\u{1b282}: 'ğ›Š‚'

CharacterDash1b283

\u{1b283}: 'ğ›Šƒ'

CharacterDash1b284

\u{1b284}: 'ğ›Š„'

CharacterDash1b285

\u{1b285}: 'ğ›Š…'

CharacterDash1b286

\u{1b286}: 'ğ›Š†'

CharacterDash1b287

\u{1b287}: 'ğ›Š‡'

CharacterDash1b288

\u{1b288}: 'ğ›Šˆ'

CharacterDash1b289

\u{1b289}: 'ğ›Š‰'

CharacterDash1b28a

\u{1b28a}: 'ğ›ŠŠ'

CharacterDash1b28b

\u{1b28b}: 'ğ›Š‹'

CharacterDash1b28c

\u{1b28c}: 'ğ›ŠŒ'

CharacterDash1b28d

\u{1b28d}: 'ğ›Š'

CharacterDash1b28e

\u{1b28e}: 'ğ›Š'

CharacterDash1b28f

\u{1b28f}: 'ğ›Š'

CharacterDash1b290

\u{1b290}: 'ğ›Š'

CharacterDash1b291

\u{1b291}: 'ğ›Š‘'

CharacterDash1b292

\u{1b292}: 'ğ›Š’'

CharacterDash1b293

\u{1b293}: 'ğ›Š“'

CharacterDash1b294

\u{1b294}: 'ğ›Š”'

CharacterDash1b295

\u{1b295}: 'ğ›Š•'

CharacterDash1b296

\u{1b296}: 'ğ›Š–'

CharacterDash1b297

\u{1b297}: 'ğ›Š—'

CharacterDash1b298

\u{1b298}: 'ğ›Š˜'

CharacterDash1b299

\u{1b299}: 'ğ›Š™'

CharacterDash1b29a

\u{1b29a}: 'ğ›Šš'

CharacterDash1b29b

\u{1b29b}: 'ğ›Š›'

CharacterDash1b29c

\u{1b29c}: 'ğ›Šœ'

CharacterDash1b29d

\u{1b29d}: 'ğ›Š'

CharacterDash1b29e

\u{1b29e}: 'ğ›Š'

CharacterDash1b29f

\u{1b29f}: 'ğ›ŠŸ'

CharacterDash1b2a0

\u{1b2a0}: 'ğ›Š '

CharacterDash1b2a1

\u{1b2a1}: 'ğ›Š¡'

CharacterDash1b2a2

\u{1b2a2}: 'ğ›Š¢'

CharacterDash1b2a3

\u{1b2a3}: 'ğ›Š£'

CharacterDash1b2a4

\u{1b2a4}: 'ğ›Š¤'

CharacterDash1b2a5

\u{1b2a5}: 'ğ›Š¥'

CharacterDash1b2a6

\u{1b2a6}: 'ğ›Š¦'

CharacterDash1b2a7

\u{1b2a7}: 'ğ›Š'

CharacterDash1b2a8

\u{1b2a8}: 'ğ›Š¨'

CharacterDash1b2a9

\u{1b2a9}: 'ğ›Š©'

CharacterDash1b2aa

\u{1b2aa}: 'ğ›Šª'

CharacterDash1b2ab

\u{1b2ab}: 'ğ›Š«'

CharacterDash1b2ac

\u{1b2ac}: 'ğ›Š¬'

CharacterDash1b2ad

\u{1b2ad}: 'ğ›Š­'

CharacterDash1b2ae

\u{1b2ae}: 'ğ›Š®'

CharacterDash1b2af

\u{1b2af}: 'ğ›Š¯'

CharacterDash1b2b0

\u{1b2b0}: 'ğ›Š°'

CharacterDash1b2b1

\u{1b2b1}: 'ğ›Š±'

CharacterDash1b2b2

\u{1b2b2}: 'ğ›Š²'

CharacterDash1b2b3

\u{1b2b3}: 'ğ›Š³'

CharacterDash1b2b4

\u{1b2b4}: 'ğ›Š´'

CharacterDash1b2b5

\u{1b2b5}: 'ğ›Šµ'

CharacterDash1b2b6

\u{1b2b6}: 'ğ›Š¶'

CharacterDash1b2b7

\u{1b2b7}: 'ğ›Š·'

CharacterDash1b2b8

\u{1b2b8}: 'ğ›Š¸'

CharacterDash1b2b9

\u{1b2b9}: 'ğ›Š¹'

CharacterDash1b2ba

\u{1b2ba}: 'ğ›Šº'

CharacterDash1b2bb

\u{1b2bb}: 'ğ›Š»'

CharacterDash1b2bc

\u{1b2bc}: 'ğ›Š¼'

CharacterDash1b2bd

\u{1b2bd}: 'ğ›Š½'

CharacterDash1b2be

\u{1b2be}: 'ğ›Š¾'

CharacterDash1b2bf

\u{1b2bf}: 'ğ›Š¿'

CharacterDash1b2c0

\u{1b2c0}: 'ğ›‹€'

CharacterDash1b2c1

\u{1b2c1}: 'ğ›‹'

CharacterDash1b2c2

\u{1b2c2}: 'ğ›‹‚'

CharacterDash1b2c3

\u{1b2c3}: 'ğ›‹ƒ'

CharacterDash1b2c4

\u{1b2c4}: 'ğ›‹„'

CharacterDash1b2c5

\u{1b2c5}: 'ğ›‹…'

CharacterDash1b2c6

\u{1b2c6}: 'ğ›‹†'

CharacterDash1b2c7

\u{1b2c7}: 'ğ›‹‡'

CharacterDash1b2c8

\u{1b2c8}: 'ğ›‹ˆ'

CharacterDash1b2c9

\u{1b2c9}: 'ğ›‹‰'

CharacterDash1b2ca

\u{1b2ca}: 'ğ›‹Š'

CharacterDash1b2cb

\u{1b2cb}: 'ğ›‹‹'

CharacterDash1b2cc

\u{1b2cc}: 'ğ›‹Œ'

CharacterDash1b2cd

\u{1b2cd}: 'ğ›‹'

CharacterDash1b2ce

\u{1b2ce}: 'ğ›‹'

CharacterDash1b2cf

\u{1b2cf}: 'ğ›‹'

CharacterDash1b2d0

\u{1b2d0}: 'ğ›‹'

CharacterDash1b2d1

\u{1b2d1}: 'ğ›‹‘'

CharacterDash1b2d2

\u{1b2d2}: 'ğ›‹’'

CharacterDash1b2d3

\u{1b2d3}: 'ğ›‹“'

CharacterDash1b2d4

\u{1b2d4}: 'ğ›‹”'

CharacterDash1b2d5

\u{1b2d5}: 'ğ›‹•'

CharacterDash1b2d6

\u{1b2d6}: 'ğ›‹–'

CharacterDash1b2d7

\u{1b2d7}: 'ğ›‹—'

CharacterDash1b2d8

\u{1b2d8}: 'ğ›‹˜'

CharacterDash1b2d9

\u{1b2d9}: 'ğ›‹™'

CharacterDash1b2da

\u{1b2da}: 'ğ›‹š'

CharacterDash1b2db

\u{1b2db}: 'ğ›‹›'

CharacterDash1b2dc

\u{1b2dc}: 'ğ›‹œ'

CharacterDash1b2dd

\u{1b2dd}: 'ğ›‹'

CharacterDash1b2de

\u{1b2de}: 'ğ›‹'

CharacterDash1b2df

\u{1b2df}: 'ğ›‹Ÿ'

CharacterDash1b2e0

\u{1b2e0}: 'ğ›‹ '

CharacterDash1b2e1

\u{1b2e1}: 'ğ›‹¡'

CharacterDash1b2e2

\u{1b2e2}: 'ğ›‹¢'

CharacterDash1b2e3

\u{1b2e3}: 'ğ›‹£'

CharacterDash1b2e4

\u{1b2e4}: 'ğ›‹¤'

CharacterDash1b2e5

\u{1b2e5}: 'ğ›‹¥'

CharacterDash1b2e6

\u{1b2e6}: 'ğ›‹¦'

CharacterDash1b2e7

\u{1b2e7}: 'ğ›‹'

CharacterDash1b2e8

\u{1b2e8}: 'ğ›‹¨'

CharacterDash1b2e9

\u{1b2e9}: 'ğ›‹©'

CharacterDash1b2ea

\u{1b2ea}: 'ğ›‹ª'

CharacterDash1b2eb

\u{1b2eb}: 'ğ›‹«'

CharacterDash1b2ec

\u{1b2ec}: 'ğ›‹¬'

CharacterDash1b2ed

\u{1b2ed}: 'ğ›‹­'

CharacterDash1b2ee

\u{1b2ee}: 'ğ›‹®'

CharacterDash1b2ef

\u{1b2ef}: 'ğ›‹¯'

CharacterDash1b2f0

\u{1b2f0}: 'ğ›‹°'

CharacterDash1b2f1

\u{1b2f1}: 'ğ›‹±'

CharacterDash1b2f2

\u{1b2f2}: 'ğ›‹²'

CharacterDash1b2f3

\u{1b2f3}: 'ğ›‹³'

CharacterDash1b2f4

\u{1b2f4}: 'ğ›‹´'

CharacterDash1b2f5

\u{1b2f5}: 'ğ›‹µ'

CharacterDash1b2f6

\u{1b2f6}: 'ğ›‹¶'

CharacterDash1b2f7

\u{1b2f7}: 'ğ›‹·'

CharacterDash1b2f8

\u{1b2f8}: 'ğ›‹¸'

CharacterDash1b2f9

\u{1b2f9}: 'ğ›‹¹'

CharacterDash1b2fa

\u{1b2fa}: 'ğ›‹º'

CharacterDash1b2fb

\u{1b2fb}: 'ğ›‹»'

Methods

impl Nushu

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case