Enum unicode_types::old_hungarian::OldHungarian

pub enum OldHungarian {
    CapitalLetterA,
    CapitalLetterAa,
    CapitalLetterEb,
    CapitalLetterAmb,
    CapitalLetterEc,
    CapitalLetterEnc,
    CapitalLetterEcs,
    CapitalLetterEd,
    CapitalLetterAnd,
    CapitalLetterE,
    CapitalLetterCloseE,
    CapitalLetterEe,
    CapitalLetterEf,
    CapitalLetterEg,
    CapitalLetterEgy,
    CapitalLetterEh,
    CapitalLetterI,
    CapitalLetterIi,
    CapitalLetterEj,
    CapitalLetterEk,
    CapitalLetterAk,
    CapitalLetterUnk,
    CapitalLetterEl,
    CapitalLetterEly,
    CapitalLetterEm,
    CapitalLetterEn,
    CapitalLetterEny,
    CapitalLetterO,
    CapitalLetterOo,
    CapitalLetterNikolsburgOe,
    CapitalLetterRudimentaOe,
    CapitalLetterOee,
    CapitalLetterEp,
    CapitalLetterEmp,
    CapitalLetterEr,
    CapitalLetterShortEr,
    CapitalLetterEs,
    CapitalLetterEsz,
    CapitalLetterEt,
    CapitalLetterEnt,
    CapitalLetterEty,
    CapitalLetterEch,
    CapitalLetterU,
    CapitalLetterUu,
    CapitalLetterNikolsburgUe,
    CapitalLetterRudimentaUe,
    CapitalLetterEv,
    CapitalLetterEz,
    CapitalLetterEzs,
    CapitalLetterEntDashShapedSign,
    CapitalLetterUs,
    SmallLetterA,
    SmallLetterAa,
    SmallLetterEb,
    SmallLetterAmb,
    SmallLetterEc,
    SmallLetterEnc,
    SmallLetterEcs,
    SmallLetterEd,
    SmallLetterAnd,
    SmallLetterE,
    SmallLetterCloseE,
    SmallLetterEe,
    SmallLetterEf,
    SmallLetterEg,
    SmallLetterEgy,
    SmallLetterEh,
    SmallLetterI,
    SmallLetterIi,
    SmallLetterEj,
    SmallLetterEk,
    SmallLetterAk,
    SmallLetterUnk,
    SmallLetterEl,
    SmallLetterEly,
    SmallLetterEm,
    SmallLetterEn,
    SmallLetterEny,
    SmallLetterO,
    SmallLetterOo,
    SmallLetterNikolsburgOe,
    SmallLetterRudimentaOe,
    SmallLetterOee,
    SmallLetterEp,
    SmallLetterEmp,
    SmallLetterEr,
    SmallLetterShortEr,
    SmallLetterEs,
    SmallLetterEsz,
    SmallLetterEt,
    SmallLetterEnt,
    SmallLetterEty,
    SmallLetterEch,
    SmallLetterU,
    SmallLetterUu,
    SmallLetterNikolsburgUe,
    SmallLetterRudimentaUe,
    SmallLetterEv,
    SmallLetterEz,
    SmallLetterEzs,
    SmallLetterEntDashShapedSign,
    SmallLetterUs,
    NumberOne,
    NumberFive,
    NumberTen,
    NumberFifty,
    NumberOneHundred,
}

An enum to represent all characters in the OldHungarian block.

Variants
CapitalLetterA

\u{10c80}: 'ğ²€'

CapitalLetterAa

\u{10c81}: 'ğ²'

CapitalLetterEb

\u{10c82}: 'ğ²‚'

CapitalLetterAmb

\u{10c83}: 'ğ²ƒ'

CapitalLetterEc

\u{10c84}: 'ğ²„'

CapitalLetterEnc

\u{10c85}: 'ğ²…'

CapitalLetterEcs

\u{10c86}: 'ğ²†'

CapitalLetterEd

\u{10c87}: 'ğ²‡'

CapitalLetterAnd

\u{10c88}: 'ğ²ˆ'

CapitalLetterE

\u{10c89}: 'ğ²‰'

CapitalLetterCloseE

\u{10c8a}: 'ğ²Š'

CapitalLetterEe

\u{10c8b}: 'ğ²‹'

CapitalLetterEf

\u{10c8c}: 'ğ²Œ'

CapitalLetterEg

\u{10c8d}: 'ğ²'

CapitalLetterEgy

\u{10c8e}: 'ğ²'

CapitalLetterEh

\u{10c8f}: 'ğ²'

CapitalLetterI

\u{10c90}: 'ğ²'

CapitalLetterIi

\u{10c91}: 'ğ²‘'

CapitalLetterEj

\u{10c92}: 'ğ²’'

CapitalLetterEk

\u{10c93}: 'ğ²“'

CapitalLetterAk

\u{10c94}: 'ğ²”'

CapitalLetterUnk

\u{10c95}: 'ğ²•'

CapitalLetterEl

\u{10c96}: 'ğ²–'

CapitalLetterEly

\u{10c97}: 'ğ²—'

CapitalLetterEm

\u{10c98}: 'ğ²˜'

CapitalLetterEn

\u{10c99}: 'ğ²™'

CapitalLetterEny

\u{10c9a}: 'ğ²š'

CapitalLetterO

\u{10c9b}: 'ğ²›'

CapitalLetterOo

\u{10c9c}: 'ğ²œ'

CapitalLetterNikolsburgOe

\u{10c9d}: 'ğ²'

CapitalLetterRudimentaOe

\u{10c9e}: 'ğ²'

CapitalLetterOee

\u{10c9f}: 'ğ²Ÿ'

CapitalLetterEp

\u{10ca0}: 'ğ² '

CapitalLetterEmp

\u{10ca1}: 'ğ²¡'

CapitalLetterEr

\u{10ca2}: 'ğ²¢'

CapitalLetterShortEr

\u{10ca3}: 'ğ²£'

CapitalLetterEs

\u{10ca4}: 'ğ²¤'

CapitalLetterEsz

\u{10ca5}: 'ğ²¥'

CapitalLetterEt

\u{10ca6}: 'ğ²¦'

CapitalLetterEnt

\u{10ca7}: 'ğ²'

CapitalLetterEty

\u{10ca8}: 'ğ²¨'

CapitalLetterEch

\u{10ca9}: 'ğ²©'

CapitalLetterU

\u{10caa}: 'ğ²ª'

CapitalLetterUu

\u{10cab}: 'ğ²«'

CapitalLetterNikolsburgUe

\u{10cac}: 'ğ²¬'

CapitalLetterRudimentaUe

\u{10cad}: 'ğ²­'

CapitalLetterEv

\u{10cae}: 'ğ²®'

CapitalLetterEz

\u{10caf}: 'ğ²¯'

CapitalLetterEzs

\u{10cb0}: 'ğ²°'

CapitalLetterEntDashShapedSign

\u{10cb1}: 'ğ²±'

CapitalLetterUs

\u{10cb2}: 'ğ²²'

SmallLetterA

\u{10cc0}: 'ğ³€'

SmallLetterAa

\u{10cc1}: 'ğ³'

SmallLetterEb

\u{10cc2}: 'ğ³‚'

SmallLetterAmb

\u{10cc3}: 'ğ³ƒ'

SmallLetterEc

\u{10cc4}: 'ğ³„'

SmallLetterEnc

\u{10cc5}: 'ğ³…'

SmallLetterEcs

\u{10cc6}: 'ğ³†'

SmallLetterEd

\u{10cc7}: 'ğ³‡'

SmallLetterAnd

\u{10cc8}: 'ğ³ˆ'

SmallLetterE

\u{10cc9}: 'ğ³‰'

SmallLetterCloseE

\u{10cca}: 'ğ³Š'

SmallLetterEe

\u{10ccb}: 'ğ³‹'

SmallLetterEf

\u{10ccc}: 'ğ³Œ'

SmallLetterEg

\u{10ccd}: 'ğ³'

SmallLetterEgy

\u{10cce}: 'ğ³'

SmallLetterEh

\u{10ccf}: 'ğ³'

SmallLetterI

\u{10cd0}: 'ğ³'

SmallLetterIi

\u{10cd1}: 'ğ³‘'

SmallLetterEj

\u{10cd2}: 'ğ³’'

SmallLetterEk

\u{10cd3}: 'ğ³“'

SmallLetterAk

\u{10cd4}: 'ğ³”'

SmallLetterUnk

\u{10cd5}: 'ğ³•'

SmallLetterEl

\u{10cd6}: 'ğ³–'

SmallLetterEly

\u{10cd7}: 'ğ³—'

SmallLetterEm

\u{10cd8}: 'ğ³˜'

SmallLetterEn

\u{10cd9}: 'ğ³™'

SmallLetterEny

\u{10cda}: 'ğ³š'

SmallLetterO

\u{10cdb}: 'ğ³›'

SmallLetterOo

\u{10cdc}: 'ğ³œ'

SmallLetterNikolsburgOe

\u{10cdd}: 'ğ³'

SmallLetterRudimentaOe

\u{10cde}: 'ğ³'

SmallLetterOee

\u{10cdf}: 'ğ³Ÿ'

SmallLetterEp

\u{10ce0}: 'ğ³ '

SmallLetterEmp

\u{10ce1}: 'ğ³¡'

SmallLetterEr

\u{10ce2}: 'ğ³¢'

SmallLetterShortEr

\u{10ce3}: 'ğ³£'

SmallLetterEs

\u{10ce4}: 'ğ³¤'

SmallLetterEsz

\u{10ce5}: 'ğ³¥'

SmallLetterEt

\u{10ce6}: 'ğ³¦'

SmallLetterEnt

\u{10ce7}: 'ğ³'

SmallLetterEty

\u{10ce8}: 'ğ³¨'

SmallLetterEch

\u{10ce9}: 'ğ³©'

SmallLetterU

\u{10cea}: 'ğ³ª'

SmallLetterUu

\u{10ceb}: 'ğ³«'

SmallLetterNikolsburgUe

\u{10cec}: 'ğ³¬'

SmallLetterRudimentaUe

\u{10ced}: 'ğ³­'

SmallLetterEv

\u{10cee}: 'ğ³®'

SmallLetterEz

\u{10cef}: 'ğ³¯'

SmallLetterEzs

\u{10cf0}: 'ğ³°'

SmallLetterEntDashShapedSign

\u{10cf1}: 'ğ³±'

SmallLetterUs

\u{10cf2}: 'ğ³²'

NumberOne

\u{10cfa}: 'ğ³º'

NumberFive

\u{10cfb}: 'ğ³»'

NumberTen

\u{10cfc}: 'ğ³¼'

NumberFifty

\u{10cfd}: 'ğ³½'

NumberOneHundred

\u{10cfe}: 'ğ³¾'

Methods

impl OldHungarian

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case