Enum unicode_types::old_persian::OldPersian

pub enum OldPersian {
    SignA,
    SignI,
    SignU,
    SignKa,
    SignKu,
    SignGa,
    SignGu,
    SignXa,
    SignCa,
    SignJa,
    SignJi,
    SignTa,
    SignTu,
    SignDa,
    SignDi,
    SignDu,
    SignTha,
    SignPa,
    SignBa,
    SignFa,
    SignNa,
    SignNu,
    SignMa,
    SignMi,
    SignMu,
    SignYa,
    SignVa,
    SignVi,
    SignRa,
    SignRu,
    SignLa,
    SignSa,
    SignZa,
    SignSha,
    SignSsa,
    SignHa,
    SignAuramazdaa,
    SignAuramazdaaDash2,
    SignAuramazdaaha,
    SignXshaayathiya,
    SignDahyaaush,
    SignDahyaaushDash2,
    SignBaga,
    SignBuumish,
    WordDivider,
    NumberOne,
    NumberTwo,
    NumberTen,
    NumberTwenty,
    NumberHundred,
}

An enum to represent all characters in the OldPersian block.

Variants
SignA

\u{103a0}: 'ğ '

SignI

\u{103a1}: 'ğ¡'

SignU

\u{103a2}: 'ğ¢'

SignKa

\u{103a3}: 'ğ£'

SignKu

\u{103a4}: 'ğ¤'

SignGa

\u{103a5}: 'ğ¥'

SignGu

\u{103a6}: 'ğ¦'

SignXa

\u{103a7}: 'ğ'

SignCa

\u{103a8}: 'ğ¨'

SignJa

\u{103a9}: 'ğ©'

SignJi

\u{103aa}: 'ğª'

SignTa

\u{103ab}: 'ğ«'

SignTu

\u{103ac}: 'ğ¬'

SignDa

\u{103ad}: 'ğ­'

SignDi

\u{103ae}: 'ğ®'

SignDu

\u{103af}: 'ğ¯'

SignTha

\u{103b0}: 'ğ°'

SignPa

\u{103b1}: 'ğ±'

SignBa

\u{103b2}: 'ğ²'

SignFa

\u{103b3}: 'ğ³'

SignNa

\u{103b4}: 'ğ´'

SignNu

\u{103b5}: 'ğµ'

SignMa

\u{103b6}: 'ğ¶'

SignMi

\u{103b7}: 'ğ·'

SignMu

\u{103b8}: 'ğ¸'

SignYa

\u{103b9}: 'ğ¹'

SignVa

\u{103ba}: 'ğº'

SignVi

\u{103bb}: 'ğ»'

SignRa

\u{103bc}: 'ğ¼'

SignRu

\u{103bd}: 'ğ½'

SignLa

\u{103be}: 'ğ¾'

SignSa

\u{103bf}: 'ğ¿'

SignZa

\u{103c0}: 'ğ€'

SignSha

\u{103c1}: 'ğ'

SignSsa

\u{103c2}: 'ğ‚'

SignHa

\u{103c3}: 'ğƒ'

SignAuramazdaa

\u{103c8}: 'ğˆ'

SignAuramazdaaDash2

\u{103c9}: 'ğ‰'

SignAuramazdaaha

\u{103ca}: 'ğŠ'

SignXshaayathiya

\u{103cb}: 'ğ‹'

SignDahyaaush

\u{103cc}: 'ğŒ'

SignDahyaaushDash2

\u{103cd}: 'ğ'

SignBaga

\u{103ce}: 'ğ'

SignBuumish

\u{103cf}: 'ğ'

WordDivider

\u{103d0}: 'ğ'

NumberOne

\u{103d1}: 'ğ‘'

NumberTwo

\u{103d2}: 'ğ’'

NumberTen

\u{103d3}: 'ğ“'

NumberTwenty

\u{103d4}: 'ğ”'

NumberHundred

\u{103d5}: 'ğ•'

Methods

impl OldPersian

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case