Enum unicode_types::old_turkic::OldTurkic

pub enum OldTurkic {
    LetterOrkhonA,
    LetterYeniseiA,
    LetterYeniseiAe,
    LetterOrkhonI,
    LetterYeniseiI,
    LetterYeniseiE,
    LetterOrkhonO,
    LetterOrkhonOe,
    LetterYeniseiOe,
    LetterOrkhonAb,
    LetterYeniseiAb,
    LetterOrkhonAeb,
    LetterYeniseiAeb,
    LetterOrkhonAg,
    LetterYeniseiAg,
    LetterOrkhonAeg,
    LetterYeniseiAeg,
    LetterOrkhonAd,
    LetterYeniseiAd,
    LetterOrkhonAed,
    LetterOrkhonEz,
    LetterYeniseiEz,
    LetterOrkhonAy,
    LetterYeniseiAy,
    LetterOrkhonAey,
    LetterYeniseiAey,
    LetterOrkhonAek,
    LetterYeniseiAek,
    LetterOrkhonOek,
    LetterYeniseiOek,
    LetterOrkhonAl,
    LetterYeniseiAl,
    LetterOrkhonAel,
    LetterOrkhonElt,
    LetterOrkhonEm,
    LetterOrkhonAn,
    LetterOrkhonAen,
    LetterYeniseiAen,
    LetterOrkhonEnt,
    LetterYeniseiEnt,
    LetterOrkhonEnc,
    LetterYeniseiEnc,
    LetterOrkhonEny,
    LetterYeniseiEny,
    LetterYeniseiAng,
    LetterOrkhonEng,
    LetterYeniseiAeng,
    LetterOrkhonEp,
    LetterOrkhonOp,
    LetterOrkhonIc,
    LetterOrkhonEc,
    LetterYeniseiEc,
    LetterOrkhonAq,
    LetterYeniseiAq,
    LetterOrkhonIq,
    LetterYeniseiIq,
    LetterOrkhonOq,
    LetterYeniseiOq,
    LetterOrkhonAr,
    LetterYeniseiAr,
    LetterOrkhonAer,
    LetterOrkhonAs,
    LetterOrkhonAes,
    LetterOrkhonAsh,
    LetterYeniseiAsh,
    LetterOrkhonEsh,
    LetterYeniseiEsh,
    LetterOrkhonAt,
    LetterYeniseiAt,
    LetterOrkhonAet,
    LetterYeniseiAet,
    LetterOrkhonOt,
    LetterOrkhonBash,
}

An enum to represent all characters in the OldTurkic block.

Variants
LetterOrkhonA

\u{10c00}: 'ğ°€'

LetterYeniseiA

\u{10c01}: 'ğ°'

LetterYeniseiAe

\u{10c02}: 'ğ°‚'

LetterOrkhonI

\u{10c03}: 'ğ°ƒ'

LetterYeniseiI

\u{10c04}: 'ğ°„'

LetterYeniseiE

\u{10c05}: 'ğ°…'

LetterOrkhonO

\u{10c06}: 'ğ°†'

LetterOrkhonOe

\u{10c07}: 'ğ°‡'

LetterYeniseiOe

\u{10c08}: 'ğ°ˆ'

LetterOrkhonAb

\u{10c09}: 'ğ°‰'

LetterYeniseiAb

\u{10c0a}: 'ğ°Š'

LetterOrkhonAeb

\u{10c0b}: 'ğ°‹'

LetterYeniseiAeb

\u{10c0c}: 'ğ°Œ'

LetterOrkhonAg

\u{10c0d}: 'ğ°'

LetterYeniseiAg

\u{10c0e}: 'ğ°'

LetterOrkhonAeg

\u{10c0f}: 'ğ°'

LetterYeniseiAeg

\u{10c10}: 'ğ°'

LetterOrkhonAd

\u{10c11}: 'ğ°‘'

LetterYeniseiAd

\u{10c12}: 'ğ°’'

LetterOrkhonAed

\u{10c13}: 'ğ°“'

LetterOrkhonEz

\u{10c14}: 'ğ°”'

LetterYeniseiEz

\u{10c15}: 'ğ°•'

LetterOrkhonAy

\u{10c16}: 'ğ°–'

LetterYeniseiAy

\u{10c17}: 'ğ°—'

LetterOrkhonAey

\u{10c18}: 'ğ°˜'

LetterYeniseiAey

\u{10c19}: 'ğ°™'

LetterOrkhonAek

\u{10c1a}: 'ğ°š'

LetterYeniseiAek

\u{10c1b}: 'ğ°›'

LetterOrkhonOek

\u{10c1c}: 'ğ°œ'

LetterYeniseiOek

\u{10c1d}: 'ğ°'

LetterOrkhonAl

\u{10c1e}: 'ğ°'

LetterYeniseiAl

\u{10c1f}: 'ğ°Ÿ'

LetterOrkhonAel

\u{10c20}: 'ğ° '

LetterOrkhonElt

\u{10c21}: 'ğ°¡'

LetterOrkhonEm

\u{10c22}: 'ğ°¢'

LetterOrkhonAn

\u{10c23}: 'ğ°£'

LetterOrkhonAen

\u{10c24}: 'ğ°¤'

LetterYeniseiAen

\u{10c25}: 'ğ°¥'

LetterOrkhonEnt

\u{10c26}: 'ğ°¦'

LetterYeniseiEnt

\u{10c27}: 'ğ°'

LetterOrkhonEnc

\u{10c28}: 'ğ°¨'

LetterYeniseiEnc

\u{10c29}: 'ğ°©'

LetterOrkhonEny

\u{10c2a}: 'ğ°ª'

LetterYeniseiEny

\u{10c2b}: 'ğ°«'

LetterYeniseiAng

\u{10c2c}: 'ğ°¬'

LetterOrkhonEng

\u{10c2d}: 'ğ°­'

LetterYeniseiAeng

\u{10c2e}: 'ğ°®'

LetterOrkhonEp

\u{10c2f}: 'ğ°¯'

LetterOrkhonOp

\u{10c30}: 'ğ°°'

LetterOrkhonIc

\u{10c31}: 'ğ°±'

LetterOrkhonEc

\u{10c32}: 'ğ°²'

LetterYeniseiEc

\u{10c33}: 'ğ°³'

LetterOrkhonAq

\u{10c34}: 'ğ°´'

LetterYeniseiAq

\u{10c35}: 'ğ°µ'

LetterOrkhonIq

\u{10c36}: 'ğ°¶'

LetterYeniseiIq

\u{10c37}: 'ğ°·'

LetterOrkhonOq

\u{10c38}: 'ğ°¸'

LetterYeniseiOq

\u{10c39}: 'ğ°¹'

LetterOrkhonAr

\u{10c3a}: 'ğ°º'

LetterYeniseiAr

\u{10c3b}: 'ğ°»'

LetterOrkhonAer

\u{10c3c}: 'ğ°¼'

LetterOrkhonAs

\u{10c3d}: 'ğ°½'

LetterOrkhonAes

\u{10c3e}: 'ğ°¾'

LetterOrkhonAsh

\u{10c3f}: 'ğ°¿'

LetterYeniseiAsh

\u{10c40}: 'ğ±€'

LetterOrkhonEsh

\u{10c41}: 'ğ±'

LetterYeniseiEsh

\u{10c42}: 'ğ±‚'

LetterOrkhonAt

\u{10c43}: 'ğ±ƒ'

LetterYeniseiAt

\u{10c44}: 'ğ±„'

LetterOrkhonAet

\u{10c45}: 'ğ±…'

LetterYeniseiAet

\u{10c46}: 'ğ±†'

LetterOrkhonOt

\u{10c47}: 'ğ±‡'

LetterOrkhonBash

\u{10c48}: 'ğ±ˆ'

Methods

impl OldTurkic

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case