Enum unicode_types::ottoman_siyaq_numbers::OttomanSiyaqNumbers

pub enum OttomanSiyaqNumbers {
    OttomanSiyaqNumberOne,
    OttomanSiyaqNumberTwo,
    OttomanSiyaqNumberThree,
    OttomanSiyaqNumberFour,
    OttomanSiyaqNumberFive,
    OttomanSiyaqNumberSix,
    OttomanSiyaqNumberSeven,
    OttomanSiyaqNumberEight,
    OttomanSiyaqNumberNine,
    OttomanSiyaqNumberTen,
    OttomanSiyaqNumberTwenty,
    OttomanSiyaqNumberThirty,
    OttomanSiyaqNumberForty,
    OttomanSiyaqNumberFifty,
    OttomanSiyaqNumberSixty,
    OttomanSiyaqNumberSeventy,
    OttomanSiyaqNumberEighty,
    OttomanSiyaqNumberNinety,
    OttomanSiyaqNumberOneHundred,
    OttomanSiyaqNumberTwoHundred,
    OttomanSiyaqNumberThreeHundred,
    OttomanSiyaqNumberFourHundred,
    OttomanSiyaqNumberFiveHundred,
    OttomanSiyaqNumberSixHundred,
    OttomanSiyaqNumberSevenHundred,
    OttomanSiyaqNumberEightHundred,
    OttomanSiyaqNumberNineHundred,
    OttomanSiyaqNumberOneThousand,
    OttomanSiyaqNumberTwoThousand,
    OttomanSiyaqNumberThreeThousand,
    OttomanSiyaqNumberFourThousand,
    OttomanSiyaqNumberFiveThousand,
    OttomanSiyaqNumberSixThousand,
    OttomanSiyaqNumberSevenThousand,
    OttomanSiyaqNumberEightThousand,
    OttomanSiyaqNumberNineThousand,
    OttomanSiyaqNumberTenThousand,
    OttomanSiyaqNumberTwentyThousand,
    OttomanSiyaqNumberThirtyThousand,
    OttomanSiyaqNumberFortyThousand,
    OttomanSiyaqNumberFiftyThousand,
    OttomanSiyaqNumberSixtyThousand,
    OttomanSiyaqNumberSeventyThousand,
    OttomanSiyaqNumberEightyThousand,
    OttomanSiyaqNumberNinetyThousand,
    OttomanSiyaqMarratan,
    OttomanSiyaqAlternateNumberTwo,
    OttomanSiyaqAlternateNumberThree,
    OttomanSiyaqAlternateNumberFour,
    OttomanSiyaqAlternateNumberFive,
    OttomanSiyaqAlternateNumberSix,
    OttomanSiyaqAlternateNumberSeven,
    OttomanSiyaqAlternateNumberEight,
    OttomanSiyaqAlternateNumberNine,
    OttomanSiyaqAlternateNumberTen,
    OttomanSiyaqAlternateNumberFourHundred,
    OttomanSiyaqAlternateNumberSixHundred,
    OttomanSiyaqAlternateNumberTwoThousand,
    OttomanSiyaqAlternateNumberTenThousand,
    OttomanSiyaqFractionOneHalf,
    OttomanSiyaqFractionOneSixth,
}

An enum to represent all characters in the OttomanSiyaqNumbers block.

Variants
OttomanSiyaqNumberOne

\u{1ed01}: 'ğ´'

OttomanSiyaqNumberTwo

\u{1ed02}: 'ğ´‚'

OttomanSiyaqNumberThree

\u{1ed03}: 'ğ´ƒ'

OttomanSiyaqNumberFour

\u{1ed04}: 'ğ´„'

OttomanSiyaqNumberFive

\u{1ed05}: 'ğ´…'

OttomanSiyaqNumberSix

\u{1ed06}: 'ğ´†'

OttomanSiyaqNumberSeven

\u{1ed07}: 'ğ´‡'

OttomanSiyaqNumberEight

\u{1ed08}: 'ğ´ˆ'

OttomanSiyaqNumberNine

\u{1ed09}: 'ğ´‰'

OttomanSiyaqNumberTen

\u{1ed0a}: 'ğ´Š'

OttomanSiyaqNumberTwenty

\u{1ed0b}: 'ğ´‹'

OttomanSiyaqNumberThirty

\u{1ed0c}: 'ğ´Œ'

OttomanSiyaqNumberForty

\u{1ed0d}: 'ğ´'

OttomanSiyaqNumberFifty

\u{1ed0e}: 'ğ´'

OttomanSiyaqNumberSixty

\u{1ed0f}: 'ğ´'

OttomanSiyaqNumberSeventy

\u{1ed10}: 'ğ´'

OttomanSiyaqNumberEighty

\u{1ed11}: 'ğ´‘'

OttomanSiyaqNumberNinety

\u{1ed12}: 'ğ´’'

OttomanSiyaqNumberOneHundred

\u{1ed13}: 'ğ´“'

OttomanSiyaqNumberTwoHundred

\u{1ed14}: 'ğ´”'

OttomanSiyaqNumberThreeHundred

\u{1ed15}: 'ğ´•'

OttomanSiyaqNumberFourHundred

\u{1ed16}: 'ğ´–'

OttomanSiyaqNumberFiveHundred

\u{1ed17}: 'ğ´—'

OttomanSiyaqNumberSixHundred

\u{1ed18}: 'ğ´˜'

OttomanSiyaqNumberSevenHundred

\u{1ed19}: 'ğ´™'

OttomanSiyaqNumberEightHundred

\u{1ed1a}: 'ğ´š'

OttomanSiyaqNumberNineHundred

\u{1ed1b}: 'ğ´›'

OttomanSiyaqNumberOneThousand

\u{1ed1c}: 'ğ´œ'

OttomanSiyaqNumberTwoThousand

\u{1ed1d}: 'ğ´'

OttomanSiyaqNumberThreeThousand

\u{1ed1e}: 'ğ´'

OttomanSiyaqNumberFourThousand

\u{1ed1f}: 'ğ´Ÿ'

OttomanSiyaqNumberFiveThousand

\u{1ed20}: 'ğ´ '

OttomanSiyaqNumberSixThousand

\u{1ed21}: 'ğ´¡'

OttomanSiyaqNumberSevenThousand

\u{1ed22}: 'ğ´¢'

OttomanSiyaqNumberEightThousand

\u{1ed23}: 'ğ´£'

OttomanSiyaqNumberNineThousand

\u{1ed24}: 'ğ´¤'

OttomanSiyaqNumberTenThousand

\u{1ed25}: 'ğ´¥'

OttomanSiyaqNumberTwentyThousand

\u{1ed26}: 'ğ´¦'

OttomanSiyaqNumberThirtyThousand

\u{1ed27}: 'ğ´'

OttomanSiyaqNumberFortyThousand

\u{1ed28}: 'ğ´¨'

OttomanSiyaqNumberFiftyThousand

\u{1ed29}: 'ğ´©'

OttomanSiyaqNumberSixtyThousand

\u{1ed2a}: 'ğ´ª'

OttomanSiyaqNumberSeventyThousand

\u{1ed2b}: 'ğ´«'

OttomanSiyaqNumberEightyThousand

\u{1ed2c}: 'ğ´¬'

OttomanSiyaqNumberNinetyThousand

\u{1ed2d}: 'ğ´­'

OttomanSiyaqMarratan

\u{1ed2e}: 'ğ´®'

OttomanSiyaqAlternateNumberTwo

\u{1ed2f}: 'ğ´¯'

OttomanSiyaqAlternateNumberThree

\u{1ed30}: 'ğ´°'

OttomanSiyaqAlternateNumberFour

\u{1ed31}: 'ğ´±'

OttomanSiyaqAlternateNumberFive

\u{1ed32}: 'ğ´²'

OttomanSiyaqAlternateNumberSix

\u{1ed33}: 'ğ´³'

OttomanSiyaqAlternateNumberSeven

\u{1ed34}: 'ğ´´'

OttomanSiyaqAlternateNumberEight

\u{1ed35}: 'ğ´µ'

OttomanSiyaqAlternateNumberNine

\u{1ed36}: 'ğ´¶'

OttomanSiyaqAlternateNumberTen

\u{1ed37}: 'ğ´·'

OttomanSiyaqAlternateNumberFourHundred

\u{1ed38}: 'ğ´¸'

OttomanSiyaqAlternateNumberSixHundred

\u{1ed39}: 'ğ´¹'

OttomanSiyaqAlternateNumberTwoThousand

\u{1ed3a}: 'ğ´º'

OttomanSiyaqAlternateNumberTenThousand

\u{1ed3b}: 'ğ´»'

OttomanSiyaqFractionOneHalf

\u{1ed3c}: 'ğ´¼'

OttomanSiyaqFractionOneSixth

\u{1ed3d}: 'ğ´½'

Methods

impl OttomanSiyaqNumbers

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case