Enum unicode_types::pahawh_hmong::PahawhHmong

pub enum PahawhHmong {
    VowelKeeb,
    VowelKeev,
    VowelKib,
    VowelKiv,
    VowelKaub,
    VowelKauv,
    VowelKub,
    VowelKuv,
    VowelKeb,
    VowelKev,
    VowelKaib,
    VowelKaiv,
    VowelKoob,
    VowelKoov,
    VowelKawb,
    VowelKawv,
    VowelKuab,
    VowelKuav,
    VowelKob,
    VowelKov,
    VowelKiab,
    VowelKiav,
    VowelKab,
    VowelKav,
    VowelKwb,
    VowelKwv,
    VowelKaab,
    VowelKaav,
    ConsonantVau,
    ConsonantNtsau,
    ConsonantLau,
    ConsonantHau,
    ConsonantNlau,
    ConsonantRau,
    ConsonantNkau,
    ConsonantQhau,
    ConsonantYau,
    ConsonantHlau,
    ConsonantMau,
    ConsonantChau,
    ConsonantNchau,
    ConsonantHnau,
    ConsonantPlhau,
    ConsonantNthau,
    ConsonantNau,
    ConsonantAu,
    ConsonantXau,
    ConsonantCau,
    MarkCimTub,
    MarkCimSo,
    MarkCimKes,
    MarkCimKhav,
    MarkCimSuam,
    MarkCimHom,
    MarkCimTaum,
    SignVosThom,
    SignVosTshabCeeb,
    SignCimCheem,
    SignVosThiab,
    SignVosFeem,
    SignXyeemNtxiv,
    SignXyeemRho,
    SignXyeemTov,
    SignXyeemFaib,
    SignVosSeev,
    SignMeejSuab,
    SignVosNrua,
    SignIbYam,
    SignXaus,
    SignCimTsovRog,
    DigitZero,
    DigitOne,
    DigitTwo,
    DigitThree,
    DigitFour,
    DigitFive,
    DigitSix,
    DigitSeven,
    DigitEight,
    DigitNine,
    NumberTens,
    NumberHundreds,
    NumberTenThousands,
    NumberMillions,
    NumberHundredMillions,
    NumberTenBillions,
    NumberTrillions,
    SignVosLub,
    SignXyoo,
    SignHli,
    SignThirdDashStageHli,
    SignZwjThaj,
    SignHnub,
    SignNqig,
    SignXiab,
    SignNtuj,
    SignAv,
    SignTxheejCeev,
    SignMeejTseeb,
    SignTau,
    SignLos,
    SignMus,
    SignCimHaisLusNtogNtog,
    SignCimCuamTshooj,
    SignCimTxwv,
    SignCimTxwvChwv,
    SignCimPubDawb,
    SignCimNresTos,
    ClanSignTsheej,
    ClanSignYeeg,
    ClanSignLis,
    ClanSignLauj,
    ClanSignXyooj,
    ClanSignKoo,
    ClanSignHawj,
    ClanSignMuas,
    ClanSignThoj,
    ClanSignTsab,
    ClanSignPhab,
    ClanSignKhab,
    ClanSignHam,
    ClanSignVaj,
    ClanSignFaj,
    ClanSignYaj,
    ClanSignTswb,
    ClanSignKwm,
}

An enum to represent all characters in the PahawhHmong block.

Variants
VowelKeeb

\u{16b00}: 'ğ–¬€'

VowelKeev

\u{16b01}: 'ğ–¬'

VowelKib

\u{16b02}: 'ğ–¬‚'

VowelKiv

\u{16b03}: 'ğ–¬ƒ'

VowelKaub

\u{16b04}: 'ğ–¬„'

VowelKauv

\u{16b05}: 'ğ–¬…'

VowelKub

\u{16b06}: 'ğ–¬†'

VowelKuv

\u{16b07}: 'ğ–¬‡'

VowelKeb

\u{16b08}: 'ğ–¬ˆ'

VowelKev

\u{16b09}: 'ğ–¬‰'

VowelKaib

\u{16b0a}: 'ğ–¬Š'

VowelKaiv

\u{16b0b}: 'ğ–¬‹'

VowelKoob

\u{16b0c}: 'ğ–¬Œ'

VowelKoov

\u{16b0d}: 'ğ–¬'

VowelKawb

\u{16b0e}: 'ğ–¬'

VowelKawv

\u{16b0f}: 'ğ–¬'

VowelKuab

\u{16b10}: 'ğ–¬'

VowelKuav

\u{16b11}: 'ğ–¬‘'

VowelKob

\u{16b12}: 'ğ–¬’'

VowelKov

\u{16b13}: 'ğ–¬“'

VowelKiab

\u{16b14}: 'ğ–¬”'

VowelKiav

\u{16b15}: 'ğ–¬•'

VowelKab

\u{16b16}: 'ğ–¬–'

VowelKav

\u{16b17}: 'ğ–¬—'

VowelKwb

\u{16b18}: 'ğ–¬˜'

VowelKwv

\u{16b19}: 'ğ–¬™'

VowelKaab

\u{16b1a}: 'ğ–¬š'

VowelKaav

\u{16b1b}: 'ğ–¬›'

ConsonantVau

\u{16b1c}: 'ğ–¬œ'

ConsonantNtsau

\u{16b1d}: 'ğ–¬'

ConsonantLau

\u{16b1e}: 'ğ–¬'

ConsonantHau

\u{16b1f}: 'ğ–¬Ÿ'

ConsonantNlau

\u{16b20}: 'ğ–¬ '

ConsonantRau

\u{16b21}: 'ğ–¬¡'

ConsonantNkau

\u{16b22}: 'ğ–¬¢'

ConsonantQhau

\u{16b23}: 'ğ–¬£'

ConsonantYau

\u{16b24}: 'ğ–¬¤'

ConsonantHlau

\u{16b25}: 'ğ–¬¥'

ConsonantMau

\u{16b26}: 'ğ–¬¦'

ConsonantChau

\u{16b27}: 'ğ–¬'

ConsonantNchau

\u{16b28}: 'ğ–¬¨'

ConsonantHnau

\u{16b29}: 'ğ–¬©'

ConsonantPlhau

\u{16b2a}: 'ğ–¬ª'

ConsonantNthau

\u{16b2b}: 'ğ–¬«'

ConsonantNau

\u{16b2c}: 'ğ–¬¬'

ConsonantAu

\u{16b2d}: 'ğ–¬­'

ConsonantXau

\u{16b2e}: 'ğ–¬®'

ConsonantCau

\u{16b2f}: 'ğ–¬¯'

MarkCimTub

\u{16b30}: 'ğ–¬°'

MarkCimSo

\u{16b31}: 'ğ–¬±'

MarkCimKes

\u{16b32}: 'ğ–¬²'

MarkCimKhav

\u{16b33}: 'ğ–¬³'

MarkCimSuam

\u{16b34}: 'ğ–¬´'

MarkCimHom

\u{16b35}: 'ğ–¬µ'

MarkCimTaum

\u{16b36}: 'ğ–¬¶'

SignVosThom

\u{16b37}: 'ğ–¬·'

SignVosTshabCeeb

\u{16b38}: 'ğ–¬¸'

SignCimCheem

\u{16b39}: 'ğ–¬¹'

SignVosThiab

\u{16b3a}: 'ğ–¬º'

SignVosFeem

\u{16b3b}: 'ğ–¬»'

SignXyeemNtxiv

\u{16b3c}: 'ğ–¬¼'

SignXyeemRho

\u{16b3d}: 'ğ–¬½'

SignXyeemTov

\u{16b3e}: 'ğ–¬¾'

SignXyeemFaib

\u{16b3f}: 'ğ–¬¿'

SignVosSeev

\u{16b40}: 'ğ–­€'

SignMeejSuab

\u{16b41}: 'ğ–­'

SignVosNrua

\u{16b42}: 'ğ–­‚'

SignIbYam

\u{16b43}: 'ğ–­ƒ'

SignXaus

\u{16b44}: 'ğ–­„'

SignCimTsovRog

\u{16b45}: 'ğ–­…'

DigitZero

\u{16b50}: 'ğ–­'

DigitOne

\u{16b51}: 'ğ–­‘'

DigitTwo

\u{16b52}: 'ğ–­’'

DigitThree

\u{16b53}: 'ğ–­“'

DigitFour

\u{16b54}: 'ğ–­”'

DigitFive

\u{16b55}: 'ğ–­•'

DigitSix

\u{16b56}: 'ğ–­–'

DigitSeven

\u{16b57}: 'ğ–­—'

DigitEight

\u{16b58}: 'ğ–­˜'

DigitNine

\u{16b59}: 'ğ–­™'

NumberTens

\u{16b5b}: 'ğ–­›'

NumberHundreds

\u{16b5c}: 'ğ–­œ'

NumberTenThousands

\u{16b5d}: 'ğ–­'

NumberMillions

\u{16b5e}: 'ğ–­'

NumberHundredMillions

\u{16b5f}: 'ğ–­Ÿ'

NumberTenBillions

\u{16b60}: 'ğ–­ '

NumberTrillions

\u{16b61}: 'ğ–­¡'

SignVosLub

\u{16b63}: 'ğ–­£'

SignXyoo

\u{16b64}: 'ğ–­¤'

SignHli

\u{16b65}: 'ğ–­¥'

SignThirdDashStageHli

\u{16b66}: 'ğ–­¦'

SignZwjThaj

\u{16b67}: 'ğ–­'

SignHnub

\u{16b68}: 'ğ–­¨'

SignNqig

\u{16b69}: 'ğ–­©'

SignXiab

\u{16b6a}: 'ğ–­ª'

SignNtuj

\u{16b6b}: 'ğ–­«'

SignAv

\u{16b6c}: 'ğ–­¬'

SignTxheejCeev

\u{16b6d}: 'ğ–­­'

SignMeejTseeb

\u{16b6e}: 'ğ–­®'

SignTau

\u{16b6f}: 'ğ–­¯'

SignLos

\u{16b70}: 'ğ–­°'

SignMus

\u{16b71}: 'ğ–­±'

SignCimHaisLusNtogNtog

\u{16b72}: 'ğ–­²'

SignCimCuamTshooj

\u{16b73}: 'ğ–­³'

SignCimTxwv

\u{16b74}: 'ğ–­´'

SignCimTxwvChwv

\u{16b75}: 'ğ–­µ'

SignCimPubDawb

\u{16b76}: 'ğ–­¶'

SignCimNresTos

\u{16b77}: 'ğ–­·'

ClanSignTsheej

\u{16b7d}: 'ğ–­½'

ClanSignYeeg

\u{16b7e}: 'ğ–­¾'

ClanSignLis

\u{16b7f}: 'ğ–­¿'

ClanSignLauj

\u{16b80}: 'ğ–®€'

ClanSignXyooj

\u{16b81}: 'ğ–®'

ClanSignKoo

\u{16b82}: 'ğ–®‚'

ClanSignHawj

\u{16b83}: 'ğ–®ƒ'

ClanSignMuas

\u{16b84}: 'ğ–®„'

ClanSignThoj

\u{16b85}: 'ğ–®…'

ClanSignTsab

\u{16b86}: 'ğ–®†'

ClanSignPhab

\u{16b87}: 'ğ–®‡'

ClanSignKhab

\u{16b88}: 'ğ–®ˆ'

ClanSignHam

\u{16b89}: 'ğ–®‰'

ClanSignVaj

\u{16b8a}: 'ğ–®Š'

ClanSignFaj

\u{16b8b}: 'ğ–®‹'

ClanSignYaj

\u{16b8c}: 'ğ–®Œ'

ClanSignTswb

\u{16b8d}: 'ğ–®'

ClanSignKwm

\u{16b8e}: 'ğ–®'

Methods

impl PahawhHmong

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case