Enum unicode_types::phonetic_extensions::PhoneticExtensions

pub enum PhoneticExtensions {
    LatinLetterSmallCapitalA,
    LatinLetterSmallCapitalAe,
    LatinSmallLetterTurnedAe,
    LatinLetterSmallCapitalBarredB,
    LatinLetterSmallCapitalC,
    LatinLetterSmallCapitalD,
    LatinLetterSmallCapitalEth,
    LatinLetterSmallCapitalE,
    LatinSmallLetterTurnedOpenE,
    LatinSmallLetterTurnedI,
    LatinLetterSmallCapitalJ,
    LatinLetterSmallCapitalK,
    LatinLetterSmallCapitalLWithStroke,
    LatinLetterSmallCapitalM,
    LatinLetterSmallCapitalReversedN,
    LatinLetterSmallCapitalO,
    LatinLetterSmallCapitalOpenO,
    LatinSmallLetterSidewaysO,
    LatinSmallLetterSidewaysOpenO,
    LatinSmallLetterSidewaysOWithStroke,
    LatinSmallLetterTurnedOe,
    LatinLetterSmallCapitalOu,
    LatinSmallLetterTopHalfO,
    LatinSmallLetterBottomHalfO,
    LatinLetterSmallCapitalP,
    LatinLetterSmallCapitalReversedR,
    LatinLetterSmallCapitalTurnedR,
    LatinLetterSmallCapitalT,
    LatinLetterSmallCapitalU,
    LatinSmallLetterSidewaysU,
    LatinSmallLetterSidewaysDiaeresizedU,
    LatinSmallLetterSidewaysTurnedM,
    LatinLetterSmallCapitalV,
    LatinLetterSmallCapitalW,
    LatinLetterSmallCapitalZ,
    LatinLetterSmallCapitalEzh,
    LatinLetterVoicedLaryngealSpirant,
    LatinLetterAin,
    GreekLetterSmallCapitalGamma,
    GreekLetterSmallCapitalLamda,
    GreekLetterSmallCapitalPi,
    GreekLetterSmallCapitalRho,
    GreekLetterSmallCapitalPsi,
    CyrillicLetterSmallCapitalEl,
    ModifierLetterCapitalA,
    ModifierLetterCapitalAe,
    ModifierLetterCapitalB,
    ModifierLetterCapitalBarredB,
    ModifierLetterCapitalD,
    ModifierLetterCapitalE,
    ModifierLetterCapitalReversedE,
    ModifierLetterCapitalG,
    ModifierLetterCapitalH,
    ModifierLetterCapitalI,
    ModifierLetterCapitalJ,
    ModifierLetterCapitalK,
    ModifierLetterCapitalL,
    ModifierLetterCapitalM,
    ModifierLetterCapitalN,
    ModifierLetterCapitalReversedN,
    ModifierLetterCapitalO,
    ModifierLetterCapitalOu,
    ModifierLetterCapitalP,
    ModifierLetterCapitalR,
    ModifierLetterCapitalT,
    ModifierLetterCapitalU,
    ModifierLetterCapitalW,
    ModifierLetterSmallA,
    ModifierLetterSmallTurnedA,
    ModifierLetterSmallAlpha,
    ModifierLetterSmallTurnedAe,
    ModifierLetterSmallB,
    ModifierLetterSmallD,
    ModifierLetterSmallE,
    ModifierLetterSmallSchwa,
    ModifierLetterSmallOpenE,
    ModifierLetterSmallTurnedOpenE,
    ModifierLetterSmallG,
    ModifierLetterSmallTurnedI,
    ModifierLetterSmallK,
    ModifierLetterSmallM,
    ModifierLetterSmallEng,
    ModifierLetterSmallO,
    ModifierLetterSmallOpenO,
    ModifierLetterSmallTopHalfO,
    ModifierLetterSmallBottomHalfO,
    ModifierLetterSmallP,
    ModifierLetterSmallT,
    ModifierLetterSmallU,
    ModifierLetterSmallSidewaysU,
    ModifierLetterSmallTurnedM,
    ModifierLetterSmallV,
    ModifierLetterSmallAin,
    ModifierLetterSmallBeta,
    ModifierLetterSmallGreekGamma,
    ModifierLetterSmallDelta,
    ModifierLetterSmallGreekPhi,
    ModifierLetterSmallChi,
    LatinSubscriptSmallLetterI,
    LatinSubscriptSmallLetterR,
    LatinSubscriptSmallLetterU,
    LatinSubscriptSmallLetterV,
    GreekSubscriptSmallLetterBeta,
    GreekSubscriptSmallLetterGamma,
    GreekSubscriptSmallLetterRho,
    GreekSubscriptSmallLetterPhi,
    GreekSubscriptSmallLetterChi,
    LatinSmallLetterUe,
    LatinSmallLetterBWithMiddleTilde,
    LatinSmallLetterDWithMiddleTilde,
    LatinSmallLetterFWithMiddleTilde,
    LatinSmallLetterMWithMiddleTilde,
    LatinSmallLetterNWithMiddleTilde,
    LatinSmallLetterPWithMiddleTilde,
    LatinSmallLetterRWithMiddleTilde,
    LatinSmallLetterRWithFishhookAndMiddleTilde,
    LatinSmallLetterSWithMiddleTilde,
    LatinSmallLetterTWithMiddleTilde,
    LatinSmallLetterZWithMiddleTilde,
    LatinSmallLetterTurnedG,
    ModifierLetterCyrillicEn,
    LatinSmallLetterInsularG,
    LatinSmallLetterThWithStrikethrough,
    LatinSmallCapitalLetterIWithStroke,
    LatinSmallLetterIotaWithStroke,
    LatinSmallLetterPWithStroke,
    LatinSmallCapitalLetterUWithStroke,
}

An enum to represent all characters in the PhoneticExtensions block.

Variants
LatinLetterSmallCapitalA

\u{1d00}: 'ᴀ'

LatinLetterSmallCapitalAe

\u{1d01}: 'ᴁ'

LatinSmallLetterTurnedAe

\u{1d02}: 'ᴂ'

LatinLetterSmallCapitalBarredB

\u{1d03}: 'ᴃ'

LatinLetterSmallCapitalC

\u{1d04}: 'ᴄ'

LatinLetterSmallCapitalD

\u{1d05}: 'ᴅ'

LatinLetterSmallCapitalEth

\u{1d06}: 'ᴆ'

LatinLetterSmallCapitalE

\u{1d07}: 'ᴇ'

LatinSmallLetterTurnedOpenE

\u{1d08}: 'ᴈ'

LatinSmallLetterTurnedI

\u{1d09}: 'ᴉ'

LatinLetterSmallCapitalJ

\u{1d0a}: 'ᴊ'

LatinLetterSmallCapitalK

\u{1d0b}: 'ᴋ'

LatinLetterSmallCapitalLWithStroke

\u{1d0c}: 'ᴌ'

LatinLetterSmallCapitalM

\u{1d0d}: 'ᴍ'

LatinLetterSmallCapitalReversedN

\u{1d0e}: 'ᴎ'

LatinLetterSmallCapitalO

\u{1d0f}: 'ᴏ'

LatinLetterSmallCapitalOpenO

\u{1d10}: 'ᴐ'

LatinSmallLetterSidewaysO

\u{1d11}: 'ᴑ'

LatinSmallLetterSidewaysOpenO

\u{1d12}: 'ᴒ'

LatinSmallLetterSidewaysOWithStroke

\u{1d13}: 'ᴓ'

LatinSmallLetterTurnedOe

\u{1d14}: 'ᴔ'

LatinLetterSmallCapitalOu

\u{1d15}: 'ᴕ'

LatinSmallLetterTopHalfO

\u{1d16}: 'ᴖ'

LatinSmallLetterBottomHalfO

\u{1d17}: 'ᴗ'

LatinLetterSmallCapitalP

\u{1d18}: 'ᴘ'

LatinLetterSmallCapitalReversedR

\u{1d19}: 'ᴙ'

LatinLetterSmallCapitalTurnedR

\u{1d1a}: 'ᴚ'

LatinLetterSmallCapitalT

\u{1d1b}: 'ᴛ'

LatinLetterSmallCapitalU

\u{1d1c}: 'ᴜ'

LatinSmallLetterSidewaysU

\u{1d1d}: 'ᴝ'

LatinSmallLetterSidewaysDiaeresizedU

\u{1d1e}: 'ᴞ'

LatinSmallLetterSidewaysTurnedM

\u{1d1f}: 'ᴟ'

LatinLetterSmallCapitalV

\u{1d20}: 'ᴠ'

LatinLetterSmallCapitalW

\u{1d21}: 'ᴡ'

LatinLetterSmallCapitalZ

\u{1d22}: 'ᴢ'

LatinLetterSmallCapitalEzh

\u{1d23}: 'ᴣ'

LatinLetterVoicedLaryngealSpirant

\u{1d24}: 'ᴤ'

LatinLetterAin

\u{1d25}: 'ᴥ'

GreekLetterSmallCapitalGamma

\u{1d26}: 'ᴦ'

GreekLetterSmallCapitalLamda

\u{1d27}: 'ᴧ'

GreekLetterSmallCapitalPi

\u{1d28}: 'ᴨ'

GreekLetterSmallCapitalRho

\u{1d29}: 'ᴩ'

GreekLetterSmallCapitalPsi

\u{1d2a}: 'ᴪ'

CyrillicLetterSmallCapitalEl

\u{1d2b}: 'ᴫ'

ModifierLetterCapitalA

\u{1d2c}: 'ᴬ'

ModifierLetterCapitalAe

\u{1d2d}: 'ᴭ'

ModifierLetterCapitalB

\u{1d2e}: 'ᴮ'

ModifierLetterCapitalBarredB

\u{1d2f}: 'ᴯ'

ModifierLetterCapitalD

\u{1d30}: 'ᴰ'

ModifierLetterCapitalE

\u{1d31}: 'ᴱ'

ModifierLetterCapitalReversedE

\u{1d32}: 'ᴲ'

ModifierLetterCapitalG

\u{1d33}: 'ᴳ'

ModifierLetterCapitalH

\u{1d34}: 'ᴴ'

ModifierLetterCapitalI

\u{1d35}: 'ᴵ'

ModifierLetterCapitalJ

\u{1d36}: 'ᴶ'

ModifierLetterCapitalK

\u{1d37}: 'ᴷ'

ModifierLetterCapitalL

\u{1d38}: 'ᴸ'

ModifierLetterCapitalM

\u{1d39}: 'ᴹ'

ModifierLetterCapitalN

\u{1d3a}: 'ᴺ'

ModifierLetterCapitalReversedN

\u{1d3b}: 'ᴻ'

ModifierLetterCapitalO

\u{1d3c}: 'ᴼ'

ModifierLetterCapitalOu

\u{1d3d}: 'ᴽ'

ModifierLetterCapitalP

\u{1d3e}: 'ᴾ'

ModifierLetterCapitalR

\u{1d3f}: 'ᴿ'

ModifierLetterCapitalT

\u{1d40}: 'ᵀ'

ModifierLetterCapitalU

\u{1d41}: 'ᵁ'

ModifierLetterCapitalW

\u{1d42}: 'ᵂ'

ModifierLetterSmallA

\u{1d43}: 'ᵃ'

ModifierLetterSmallTurnedA

\u{1d44}: 'ᵄ'

ModifierLetterSmallAlpha

\u{1d45}: 'ᵅ'

ModifierLetterSmallTurnedAe

\u{1d46}: 'ᵆ'

ModifierLetterSmallB

\u{1d47}: 'ᵇ'

ModifierLetterSmallD

\u{1d48}: 'ᵈ'

ModifierLetterSmallE

\u{1d49}: 'ᵉ'

ModifierLetterSmallSchwa

\u{1d4a}: 'ᵊ'

ModifierLetterSmallOpenE

\u{1d4b}: 'ᵋ'

ModifierLetterSmallTurnedOpenE

\u{1d4c}: 'ᵌ'

ModifierLetterSmallG

\u{1d4d}: 'ᵍ'

ModifierLetterSmallTurnedI

\u{1d4e}: 'ᵎ'

ModifierLetterSmallK

\u{1d4f}: 'ᵏ'

ModifierLetterSmallM

\u{1d50}: 'ᵐ'

ModifierLetterSmallEng

\u{1d51}: 'ᵑ'

ModifierLetterSmallO

\u{1d52}: 'ᵒ'

ModifierLetterSmallOpenO

\u{1d53}: 'ᵓ'

ModifierLetterSmallTopHalfO

\u{1d54}: 'ᵔ'

ModifierLetterSmallBottomHalfO

\u{1d55}: 'ᵕ'

ModifierLetterSmallP

\u{1d56}: 'ᵖ'

ModifierLetterSmallT

\u{1d57}: 'ᵗ'

ModifierLetterSmallU

\u{1d58}: 'ᵘ'

ModifierLetterSmallSidewaysU

\u{1d59}: 'ᵙ'

ModifierLetterSmallTurnedM

\u{1d5a}: 'ᵚ'

ModifierLetterSmallV

\u{1d5b}: 'ᵛ'

ModifierLetterSmallAin

\u{1d5c}: 'ᵜ'

ModifierLetterSmallBeta

\u{1d5d}: 'ᵝ'

ModifierLetterSmallGreekGamma

\u{1d5e}: 'ᵞ'

ModifierLetterSmallDelta

\u{1d5f}: 'ᵟ'

ModifierLetterSmallGreekPhi

\u{1d60}: 'ᵠ'

ModifierLetterSmallChi

\u{1d61}: 'ᵡ'

LatinSubscriptSmallLetterI

\u{1d62}: 'ᵢ'

LatinSubscriptSmallLetterR

\u{1d63}: 'ᵣ'

LatinSubscriptSmallLetterU

\u{1d64}: 'ᵤ'

LatinSubscriptSmallLetterV

\u{1d65}: 'ᵥ'

GreekSubscriptSmallLetterBeta

\u{1d66}: 'ᵦ'

GreekSubscriptSmallLetterGamma

\u{1d67}: 'ᵧ'

GreekSubscriptSmallLetterRho

\u{1d68}: 'ᵨ'

GreekSubscriptSmallLetterPhi

\u{1d69}: 'ᵩ'

GreekSubscriptSmallLetterChi

\u{1d6a}: 'ᵪ'

LatinSmallLetterUe

\u{1d6b}: 'ᵫ'

LatinSmallLetterBWithMiddleTilde

\u{1d6c}: 'ᵬ'

LatinSmallLetterDWithMiddleTilde

\u{1d6d}: 'ᵭ'

LatinSmallLetterFWithMiddleTilde

\u{1d6e}: 'ᵮ'

LatinSmallLetterMWithMiddleTilde

\u{1d6f}: 'ᵯ'

LatinSmallLetterNWithMiddleTilde

\u{1d70}: 'ᵰ'

LatinSmallLetterPWithMiddleTilde

\u{1d71}: 'ᵱ'

LatinSmallLetterRWithMiddleTilde

\u{1d72}: 'ᵲ'

LatinSmallLetterRWithFishhookAndMiddleTilde

\u{1d73}: 'ᵳ'

LatinSmallLetterSWithMiddleTilde

\u{1d74}: 'ᵴ'

LatinSmallLetterTWithMiddleTilde

\u{1d75}: 'ᵵ'

LatinSmallLetterZWithMiddleTilde

\u{1d76}: 'ᵶ'

LatinSmallLetterTurnedG

\u{1d77}: 'ᵷ'

ModifierLetterCyrillicEn

\u{1d78}: 'ᵸ'

LatinSmallLetterInsularG

\u{1d79}: 'ᵹ'

LatinSmallLetterThWithStrikethrough

\u{1d7a}: 'ᵺ'

LatinSmallCapitalLetterIWithStroke

\u{1d7b}: 'ᵻ'

LatinSmallLetterIotaWithStroke

\u{1d7c}: 'ᵼ'

LatinSmallLetterPWithStroke

\u{1d7d}: 'ᵽ'

LatinSmallCapitalLetterUWithStroke

\u{1d7e}: 'ᵾ'

Methods

impl PhoneticExtensions

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case