Enum unicode_types::playing_cards::PlayingCards

pub enum PlayingCards {
    PlayingCardBack,
    PlayingCardAceOfSpades,
    PlayingCardTwoOfSpades,
    PlayingCardThreeOfSpades,
    PlayingCardFourOfSpades,
    PlayingCardFiveOfSpades,
    PlayingCardSixOfSpades,
    PlayingCardSevenOfSpades,
    PlayingCardEightOfSpades,
    PlayingCardNineOfSpades,
    PlayingCardTenOfSpades,
    PlayingCardJackOfSpades,
    PlayingCardKnightOfSpades,
    PlayingCardQueenOfSpades,
    PlayingCardKingOfSpades,
    PlayingCardAceOfHearts,
    PlayingCardTwoOfHearts,
    PlayingCardThreeOfHearts,
    PlayingCardFourOfHearts,
    PlayingCardFiveOfHearts,
    PlayingCardSixOfHearts,
    PlayingCardSevenOfHearts,
    PlayingCardEightOfHearts,
    PlayingCardNineOfHearts,
    PlayingCardTenOfHearts,
    PlayingCardJackOfHearts,
    PlayingCardKnightOfHearts,
    PlayingCardQueenOfHearts,
    PlayingCardKingOfHearts,
    PlayingCardRedJoker,
    PlayingCardAceOfDiamonds,
    PlayingCardTwoOfDiamonds,
    PlayingCardThreeOfDiamonds,
    PlayingCardFourOfDiamonds,
    PlayingCardFiveOfDiamonds,
    PlayingCardSixOfDiamonds,
    PlayingCardSevenOfDiamonds,
    PlayingCardEightOfDiamonds,
    PlayingCardNineOfDiamonds,
    PlayingCardTenOfDiamonds,
    PlayingCardJackOfDiamonds,
    PlayingCardKnightOfDiamonds,
    PlayingCardQueenOfDiamonds,
    PlayingCardKingOfDiamonds,
    PlayingCardBlackJoker,
    PlayingCardAceOfClubs,
    PlayingCardTwoOfClubs,
    PlayingCardThreeOfClubs,
    PlayingCardFourOfClubs,
    PlayingCardFiveOfClubs,
    PlayingCardSixOfClubs,
    PlayingCardSevenOfClubs,
    PlayingCardEightOfClubs,
    PlayingCardNineOfClubs,
    PlayingCardTenOfClubs,
    PlayingCardJackOfClubs,
    PlayingCardKnightOfClubs,
    PlayingCardQueenOfClubs,
    PlayingCardKingOfClubs,
    PlayingCardWhiteJoker,
    PlayingCardFool,
    PlayingCardTrumpDash1,
    PlayingCardTrumpDash2,
    PlayingCardTrumpDash3,
    PlayingCardTrumpDash4,
    PlayingCardTrumpDash5,
    PlayingCardTrumpDash6,
    PlayingCardTrumpDash7,
    PlayingCardTrumpDash8,
    PlayingCardTrumpDash9,
    PlayingCardTrumpDash10,
    PlayingCardTrumpDash11,
    PlayingCardTrumpDash12,
    PlayingCardTrumpDash13,
    PlayingCardTrumpDash14,
    PlayingCardTrumpDash15,
    PlayingCardTrumpDash16,
    PlayingCardTrumpDash17,
    PlayingCardTrumpDash18,
    PlayingCardTrumpDash19,
    PlayingCardTrumpDash20,
    PlayingCardTrumpDash21,
}

An enum to represent all characters in the PlayingCards block.

Variants
PlayingCardBack

\u{1f0a0}: 'ğŸ‚ '

PlayingCardAceOfSpades

\u{1f0a1}: 'ğŸ‚¡'

PlayingCardTwoOfSpades

\u{1f0a2}: 'ğŸ‚¢'

PlayingCardThreeOfSpades

\u{1f0a3}: 'ğŸ‚£'

PlayingCardFourOfSpades

\u{1f0a4}: 'ğŸ‚¤'

PlayingCardFiveOfSpades

\u{1f0a5}: 'ğŸ‚¥'

PlayingCardSixOfSpades

\u{1f0a6}: 'ğŸ‚¦'

PlayingCardSevenOfSpades

\u{1f0a7}: 'ğŸ‚'

PlayingCardEightOfSpades

\u{1f0a8}: 'ğŸ‚¨'

PlayingCardNineOfSpades

\u{1f0a9}: 'ğŸ‚©'

PlayingCardTenOfSpades

\u{1f0aa}: 'ğŸ‚ª'

PlayingCardJackOfSpades

\u{1f0ab}: 'ğŸ‚«'

PlayingCardKnightOfSpades

\u{1f0ac}: 'ğŸ‚¬'

PlayingCardQueenOfSpades

\u{1f0ad}: 'ğŸ‚­'

PlayingCardKingOfSpades

\u{1f0ae}: 'ğŸ‚®'

PlayingCardAceOfHearts

\u{1f0b1}: 'ğŸ‚±'

PlayingCardTwoOfHearts

\u{1f0b2}: 'ğŸ‚²'

PlayingCardThreeOfHearts

\u{1f0b3}: 'ğŸ‚³'

PlayingCardFourOfHearts

\u{1f0b4}: 'ğŸ‚´'

PlayingCardFiveOfHearts

\u{1f0b5}: 'ğŸ‚µ'

PlayingCardSixOfHearts

\u{1f0b6}: 'ğŸ‚¶'

PlayingCardSevenOfHearts

\u{1f0b7}: 'ğŸ‚·'

PlayingCardEightOfHearts

\u{1f0b8}: 'ğŸ‚¸'

PlayingCardNineOfHearts

\u{1f0b9}: 'ğŸ‚¹'

PlayingCardTenOfHearts

\u{1f0ba}: 'ğŸ‚º'

PlayingCardJackOfHearts

\u{1f0bb}: 'ğŸ‚»'

PlayingCardKnightOfHearts

\u{1f0bc}: 'ğŸ‚¼'

PlayingCardQueenOfHearts

\u{1f0bd}: 'ğŸ‚½'

PlayingCardKingOfHearts

\u{1f0be}: 'ğŸ‚¾'

PlayingCardRedJoker

\u{1f0bf}: 'ğŸ‚¿'

PlayingCardAceOfDiamonds

\u{1f0c1}: 'ğŸƒ'

PlayingCardTwoOfDiamonds

\u{1f0c2}: 'ğŸƒ‚'

PlayingCardThreeOfDiamonds

\u{1f0c3}: 'ğŸƒƒ'

PlayingCardFourOfDiamonds

\u{1f0c4}: 'ğŸƒ„'

PlayingCardFiveOfDiamonds

\u{1f0c5}: 'ğŸƒ…'

PlayingCardSixOfDiamonds

\u{1f0c6}: 'ğŸƒ†'

PlayingCardSevenOfDiamonds

\u{1f0c7}: 'ğŸƒ‡'

PlayingCardEightOfDiamonds

\u{1f0c8}: 'ğŸƒˆ'

PlayingCardNineOfDiamonds

\u{1f0c9}: 'ğŸƒ‰'

PlayingCardTenOfDiamonds

\u{1f0ca}: 'ğŸƒŠ'

PlayingCardJackOfDiamonds

\u{1f0cb}: 'ğŸƒ‹'

PlayingCardKnightOfDiamonds

\u{1f0cc}: 'ğŸƒŒ'

PlayingCardQueenOfDiamonds

\u{1f0cd}: 'ğŸƒ'

PlayingCardKingOfDiamonds

\u{1f0ce}: 'ğŸƒ'

PlayingCardBlackJoker

\u{1f0cf}: 'ğŸƒ'

PlayingCardAceOfClubs

\u{1f0d1}: 'ğŸƒ‘'

PlayingCardTwoOfClubs

\u{1f0d2}: 'ğŸƒ’'

PlayingCardThreeOfClubs

\u{1f0d3}: 'ğŸƒ“'

PlayingCardFourOfClubs

\u{1f0d4}: 'ğŸƒ”'

PlayingCardFiveOfClubs

\u{1f0d5}: 'ğŸƒ•'

PlayingCardSixOfClubs

\u{1f0d6}: 'ğŸƒ–'

PlayingCardSevenOfClubs

\u{1f0d7}: 'ğŸƒ—'

PlayingCardEightOfClubs

\u{1f0d8}: 'ğŸƒ˜'

PlayingCardNineOfClubs

\u{1f0d9}: 'ğŸƒ™'

PlayingCardTenOfClubs

\u{1f0da}: 'ğŸƒš'

PlayingCardJackOfClubs

\u{1f0db}: 'ğŸƒ›'

PlayingCardKnightOfClubs

\u{1f0dc}: 'ğŸƒœ'

PlayingCardQueenOfClubs

\u{1f0dd}: 'ğŸƒ'

PlayingCardKingOfClubs

\u{1f0de}: 'ğŸƒ'

PlayingCardWhiteJoker

\u{1f0df}: 'ğŸƒŸ'

PlayingCardFool

\u{1f0e0}: 'ğŸƒ '

PlayingCardTrumpDash1

\u{1f0e1}: 'ğŸƒ¡'

PlayingCardTrumpDash2

\u{1f0e2}: 'ğŸƒ¢'

PlayingCardTrumpDash3

\u{1f0e3}: 'ğŸƒ£'

PlayingCardTrumpDash4

\u{1f0e4}: 'ğŸƒ¤'

PlayingCardTrumpDash5

\u{1f0e5}: 'ğŸƒ¥'

PlayingCardTrumpDash6

\u{1f0e6}: 'ğŸƒ¦'

PlayingCardTrumpDash7

\u{1f0e7}: 'ğŸƒ'

PlayingCardTrumpDash8

\u{1f0e8}: 'ğŸƒ¨'

PlayingCardTrumpDash9

\u{1f0e9}: 'ğŸƒ©'

PlayingCardTrumpDash10

\u{1f0ea}: 'ğŸƒª'

PlayingCardTrumpDash11

\u{1f0eb}: 'ğŸƒ«'

PlayingCardTrumpDash12

\u{1f0ec}: 'ğŸƒ¬'

PlayingCardTrumpDash13

\u{1f0ed}: 'ğŸƒ­'

PlayingCardTrumpDash14

\u{1f0ee}: 'ğŸƒ®'

PlayingCardTrumpDash15

\u{1f0ef}: 'ğŸƒ¯'

PlayingCardTrumpDash16

\u{1f0f0}: 'ğŸƒ°'

PlayingCardTrumpDash17

\u{1f0f1}: 'ğŸƒ±'

PlayingCardTrumpDash18

\u{1f0f2}: 'ğŸƒ²'

PlayingCardTrumpDash19

\u{1f0f3}: 'ğŸƒ³'

PlayingCardTrumpDash20

\u{1f0f4}: 'ğŸƒ´'

PlayingCardTrumpDash21

\u{1f0f5}: 'ğŸƒµ'

Methods

impl PlayingCards

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case