Enum unicode_types::runic::Runic

pub enum Runic {
    LetterFehuFeohFeF,
    LetterV,
    LetterUruzUrU,
    LetterYr,
    LetterY,
    LetterW,
    LetterThurisazThursThorn,
    LetterEth,
    LetterAnsuzA,
    LetterOsO,
    LetterAcA,
    LetterAesc,
    LetterLongDashBranchDashOssO,
    LetterShortDashTwigDashOssO,
    LetterO,
    LetterOe,
    LetterOn,
    LetterRaidoRadReidR,
    LetterKauna,
    LetterCen,
    LetterKaunK,
    LetterG,
    LetterEng,
    LetterGeboGyfuG,
    LetterGar,
    LetterWunjoWynnW,
    LetterHaglazH,
    LetterHaeglH,
    LetterLongDashBranchDashHagallH,
    LetterShortDashTwigDashHagallH,
    LetterNaudizNydNaudN,
    LetterShortDashTwigDashNaudN,
    LetterDottedDashN,
    LetterIsazIsIssI,
    LetterE,
    LetterJeranJ,
    LetterGer,
    LetterLongDashBranchDashArAe,
    LetterShortDashTwigDashArA,
    LetterIwazEoh,
    LetterPerthoPeorthP,
    LetterAlgizEolhx,
    LetterSowiloS,
    LetterSigelLongDashBranchDashSolS,
    LetterShortDashTwigDashSolS,
    LetterC,
    LetterZ,
    LetterTiwazTirTyrT,
    LetterShortDashTwigDashTyrT,
    LetterD,
    LetterBerkananBeorcBjarkanB,
    LetterShortDashTwigDashBjarkanB,
    LetterDottedDashP,
    LetterOpenDashP,
    LetterEhwazEhE,
    LetterMannazManM,
    LetterLongDashBranchDashMadrM,
    LetterShortDashTwigDashMadrM,
    LetterLaukazLaguLogrL,
    LetterDottedDashL,
    LetterIngwaz,
    LetterIng,
    LetterDagazDaegD,
    LetterOthalanEthelO,
    LetterEar,
    LetterIor,
    LetterCweorth,
    LetterCalc,
    LetterCealc,
    LetterStan,
    LetterLongDashBranchDashYr,
    LetterShortDashTwigDashYr,
    LetterIcelandicDashYr,
    LetterQ,
    LetterX,
    SinglePunctuation,
    MultiplePunctuation,
    CrossPunctuation,
    ArlaugSymbol,
    TvimadurSymbol,
    BelgthorSymbol,
    LetterK,
    LetterSh,
    LetterOo,
    LetterFranksCasketOs,
    LetterFranksCasketIs,
    LetterFranksCasketEh,
    LetterFranksCasketAc,
    LetterFranksCasketAesc,
}

An enum to represent all characters in the Runic block.

Variants
LetterFehuFeohFeF

\u{16a0}: 'ᚠ'

LetterV

\u{16a1}: 'ᚡ'

LetterUruzUrU

\u{16a2}: 'ᚢ'

LetterYr

\u{16a3}: 'ᚣ'

LetterY

\u{16a4}: 'ᚤ'

LetterW

\u{16a5}: 'ᚥ'

LetterThurisazThursThorn

\u{16a6}: 'ᚦ'

LetterEth

\u{16a7}: 'ᚧ'

LetterAnsuzA

\u{16a8}: 'ᚨ'

LetterOsO

\u{16a9}: 'ᚩ'

LetterAcA

\u{16aa}: 'ᚪ'

LetterAesc

\u{16ab}: 'ᚫ'

LetterLongDashBranchDashOssO

\u{16ac}: 'ᚬ'

LetterShortDashTwigDashOssO

\u{16ad}: 'ᚭ'

LetterO

\u{16ae}: 'ᚮ'

LetterOe

\u{16af}: 'ᚯ'

LetterOn

\u{16b0}: 'ᚰ'

LetterRaidoRadReidR

\u{16b1}: 'ᚱ'

LetterKauna

\u{16b2}: 'ᚲ'

LetterCen

\u{16b3}: 'ᚳ'

LetterKaunK

\u{16b4}: 'ᚴ'

LetterG

\u{16b5}: 'ᚵ'

LetterEng

\u{16b6}: 'ᚶ'

LetterGeboGyfuG

\u{16b7}: 'ᚷ'

LetterGar

\u{16b8}: 'ᚸ'

LetterWunjoWynnW

\u{16b9}: 'ᚹ'

LetterHaglazH

\u{16ba}: 'ᚺ'

LetterHaeglH

\u{16bb}: 'ᚻ'

LetterLongDashBranchDashHagallH

\u{16bc}: 'ᚼ'

LetterShortDashTwigDashHagallH

\u{16bd}: 'ᚽ'

LetterNaudizNydNaudN

\u{16be}: 'ᚾ'

LetterShortDashTwigDashNaudN

\u{16bf}: 'ᚿ'

LetterDottedDashN

\u{16c0}: 'ᛀ'

LetterIsazIsIssI

\u{16c1}: 'ᛁ'

LetterE

\u{16c2}: 'ᛂ'

LetterJeranJ

\u{16c3}: 'ᛃ'

LetterGer

\u{16c4}: 'ᛄ'

LetterLongDashBranchDashArAe

\u{16c5}: 'ᛅ'

LetterShortDashTwigDashArA

\u{16c6}: 'ᛆ'

LetterIwazEoh

\u{16c7}: 'ᛇ'

LetterPerthoPeorthP

\u{16c8}: 'ᛈ'

LetterAlgizEolhx

\u{16c9}: 'ᛉ'

LetterSowiloS

\u{16ca}: 'ᛊ'

LetterSigelLongDashBranchDashSolS

\u{16cb}: 'ᛋ'

LetterShortDashTwigDashSolS

\u{16cc}: 'ᛌ'

LetterC

\u{16cd}: 'ᛍ'

LetterZ

\u{16ce}: 'ᛎ'

LetterTiwazTirTyrT

\u{16cf}: 'ᛏ'

LetterShortDashTwigDashTyrT

\u{16d0}: 'ᛐ'

LetterD

\u{16d1}: 'ᛑ'

LetterBerkananBeorcBjarkanB

\u{16d2}: 'ᛒ'

LetterShortDashTwigDashBjarkanB

\u{16d3}: 'ᛓ'

LetterDottedDashP

\u{16d4}: 'ᛔ'

LetterOpenDashP

\u{16d5}: 'ᛕ'

LetterEhwazEhE

\u{16d6}: 'ᛖ'

LetterMannazManM

\u{16d7}: 'ᛗ'

LetterLongDashBranchDashMadrM

\u{16d8}: 'ᛘ'

LetterShortDashTwigDashMadrM

\u{16d9}: 'ᛙ'

LetterLaukazLaguLogrL

\u{16da}: 'ᛚ'

LetterDottedDashL

\u{16db}: 'ᛛ'

LetterIngwaz

\u{16dc}: 'ᛜ'

LetterIng

\u{16dd}: 'ᛝ'

LetterDagazDaegD

\u{16de}: 'ᛞ'

LetterOthalanEthelO

\u{16df}: 'ᛟ'

LetterEar

\u{16e0}: 'ᛠ'

LetterIor

\u{16e1}: 'ᛡ'

LetterCweorth

\u{16e2}: 'ᛢ'

LetterCalc

\u{16e3}: 'ᛣ'

LetterCealc

\u{16e4}: 'ᛤ'

LetterStan

\u{16e5}: 'ᛥ'

LetterLongDashBranchDashYr

\u{16e6}: 'ᛦ'

LetterShortDashTwigDashYr

\u{16e7}: 'ᛧ'

LetterIcelandicDashYr

\u{16e8}: 'ᛨ'

LetterQ

\u{16e9}: 'ᛩ'

LetterX

\u{16ea}: 'ᛪ'

SinglePunctuation

\u{16eb}: '᛫'

MultiplePunctuation

\u{16ec}: '᛬'

CrossPunctuation

\u{16ed}: '᛭'

ArlaugSymbol

\u{16ee}: 'ᛮ'

TvimadurSymbol

\u{16ef}: 'ᛯ'

BelgthorSymbol

\u{16f0}: 'ᛰ'

LetterK

\u{16f1}: 'ᛱ'

LetterSh

\u{16f2}: 'ᛲ'

LetterOo

\u{16f3}: 'ᛳ'

LetterFranksCasketOs

\u{16f4}: 'ᛴ'

LetterFranksCasketIs

\u{16f5}: 'ᛵ'

LetterFranksCasketEh

\u{16f6}: 'ᛶ'

LetterFranksCasketAc

\u{16f7}: 'ᛷ'

LetterFranksCasketAesc

\u{16f8}: 'ᛸ'

Methods

impl Runic

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case