Enum unicode_types::sinhala::Sinhala

pub enum Sinhala {
    SignAnusvaraya,
    SignVisargaya,
    LetterAyanna,
    LetterAayanna,
    LetterAeyanna,
    LetterAeeyanna,
    LetterIyanna,
    LetterIiyanna,
    LetterUyanna,
    LetterUuyanna,
    LetterIruyanna,
    LetterIruuyanna,
    LetterIluyanna,
    LetterIluuyanna,
    LetterEyanna,
    LetterEeyanna,
    LetterAiyanna,
    LetterOyanna,
    LetterOoyanna,
    LetterAuyanna,
    LetterAlpapraanaKayanna,
    LetterMahaapraanaKayanna,
    LetterAlpapraanaGayanna,
    LetterMahaapraanaGayanna,
    LetterKantajaNaasikyaya,
    LetterSanyakaGayanna,
    LetterAlpapraanaCayanna,
    LetterMahaapraanaCayanna,
    LetterAlpapraanaJayanna,
    LetterMahaapraanaJayanna,
    LetterTaalujaNaasikyaya,
    LetterTaalujaSanyoogaNaaksikyaya,
    LetterSanyakaJayanna,
    LetterAlpapraanaTtayanna,
    LetterMahaapraanaTtayanna,
    LetterAlpapraanaDdayanna,
    LetterMahaapraanaDdayanna,
    LetterMuurdhajaNayanna,
    LetterSanyakaDdayanna,
    LetterAlpapraanaTayanna,
    LetterMahaapraanaTayanna,
    LetterAlpapraanaDayanna,
    LetterMahaapraanaDayanna,
    LetterDantajaNayanna,
    LetterSanyakaDayanna,
    LetterAlpapraanaPayanna,
    LetterMahaapraanaPayanna,
    LetterAlpapraanaBayanna,
    LetterMahaapraanaBayanna,
    LetterMayanna,
    LetterAmbaBayanna,
    LetterYayanna,
    LetterRayanna,
    LetterDantajaLayanna,
    LetterVayanna,
    LetterTaalujaSayanna,
    LetterMuurdhajaSayanna,
    LetterDantajaSayanna,
    LetterHayanna,
    LetterMuurdhajaLayanna,
    LetterFayanna,
    SignAlDashLakuna,
    VowelSignAelaDashPilla,
    VowelSignKettiAedaDashPilla,
    VowelSignDigaAedaDashPilla,
    VowelSignKettiIsDashPilla,
    VowelSignDigaIsDashPilla,
    VowelSignKettiPaaDashPilla,
    VowelSignDigaPaaDashPilla,
    VowelSignGaettaDashPilla,
    VowelSignKombuva,
    VowelSignDigaKombuva,
    VowelSignKombuDeka,
    VowelSignKombuvaHaaAelaDashPilla,
    VowelSignKombuvaHaaDigaAelaDashPilla,
    VowelSignKombuvaHaaGayanukitta,
    VowelSignGayanukitta,
    LithDigitZero,
    LithDigitOne,
    LithDigitTwo,
    LithDigitThree,
    LithDigitFour,
    LithDigitFive,
    LithDigitSix,
    LithDigitSeven,
    LithDigitEight,
    LithDigitNine,
    VowelSignDigaGaettaDashPilla,
    VowelSignDigaGayanukitta,
    PunctuationKunddaliya,
}

An enum to represent all characters in the Sinhala block.

Variants
SignAnusvaraya

\u{d82}: 'ං'

SignVisargaya

\u{d83}: 'ඃ'

LetterAyanna

\u{d85}: 'අ'

LetterAayanna

\u{d86}: 'ආ'

LetterAeyanna

\u{d87}: 'ඇ'

LetterAeeyanna

\u{d88}: 'ඈ'

LetterIyanna

\u{d89}: 'ඉ'

LetterIiyanna

\u{d8a}: 'ඊ'

LetterUyanna

\u{d8b}: 'උ'

LetterUuyanna

\u{d8c}: 'ඌ'

LetterIruyanna

\u{d8d}: 'ඍ'

LetterIruuyanna

\u{d8e}: 'ඎ'

LetterIluyanna

\u{d8f}: 'ඏ'

LetterIluuyanna

\u{d90}: 'ඐ'

LetterEyanna

\u{d91}: 'එ'

LetterEeyanna

\u{d92}: 'ඒ'

LetterAiyanna

\u{d93}: 'ඓ'

LetterOyanna

\u{d94}: 'ඔ'

LetterOoyanna

\u{d95}: 'ඕ'

LetterAuyanna

\u{d96}: 'ඖ'

LetterAlpapraanaKayanna

\u{d9a}: 'ක'

LetterMahaapraanaKayanna

\u{d9b}: 'ඛ'

LetterAlpapraanaGayanna

\u{d9c}: 'ග'

LetterMahaapraanaGayanna

\u{d9d}: 'ඝ'

LetterKantajaNaasikyaya

\u{d9e}: 'ඞ'

LetterSanyakaGayanna

\u{d9f}: 'ඟ'

LetterAlpapraanaCayanna

\u{da0}: 'ච'

LetterMahaapraanaCayanna

\u{da1}: 'ඡ'

LetterAlpapraanaJayanna

\u{da2}: 'ජ'

LetterMahaapraanaJayanna

\u{da3}: 'ඣ'

LetterTaalujaNaasikyaya

\u{da4}: 'ඤ'

LetterTaalujaSanyoogaNaaksikyaya

\u{da5}: 'ඥ'

LetterSanyakaJayanna

\u{da6}: 'ඦ'

LetterAlpapraanaTtayanna

\u{da7}: 'ට'

LetterMahaapraanaTtayanna

\u{da8}: 'ඨ'

LetterAlpapraanaDdayanna

\u{da9}: 'ඩ'

LetterMahaapraanaDdayanna

\u{daa}: 'ඪ'

LetterMuurdhajaNayanna

\u{dab}: 'ණ'

LetterSanyakaDdayanna

\u{dac}: 'ඬ'

LetterAlpapraanaTayanna

\u{dad}: 'ත'

LetterMahaapraanaTayanna

\u{dae}: 'ථ'

LetterAlpapraanaDayanna

\u{daf}: 'ද'

LetterMahaapraanaDayanna

\u{db0}: 'ධ'

LetterDantajaNayanna

\u{db1}: 'න'

LetterSanyakaDayanna

\u{db3}: 'ඳ'

LetterAlpapraanaPayanna

\u{db4}: 'ප'

LetterMahaapraanaPayanna

\u{db5}: 'ඵ'

LetterAlpapraanaBayanna

\u{db6}: 'බ'

LetterMahaapraanaBayanna

\u{db7}: 'භ'

LetterMayanna

\u{db8}: 'ම'

LetterAmbaBayanna

\u{db9}: 'ඹ'

LetterYayanna

\u{dba}: 'ය'

LetterRayanna

\u{dbb}: 'ර'

LetterDantajaLayanna

\u{dbd}: 'ල'

LetterVayanna

\u{dc0}: 'ව'

LetterTaalujaSayanna

\u{dc1}: 'ශ'

LetterMuurdhajaSayanna

\u{dc2}: 'ෂ'

LetterDantajaSayanna

\u{dc3}: 'ස'

LetterHayanna

\u{dc4}: 'හ'

LetterMuurdhajaLayanna

\u{dc5}: 'ළ'

LetterFayanna

\u{dc6}: 'ෆ'

SignAlDashLakuna

\u{dca}: '්'

VowelSignAelaDashPilla

\u{dcf}: 'ා'

VowelSignKettiAedaDashPilla

\u{dd0}: 'ැ'

VowelSignDigaAedaDashPilla

\u{dd1}: 'ෑ'

VowelSignKettiIsDashPilla

\u{dd2}: 'ි'

VowelSignDigaIsDashPilla

\u{dd3}: 'ී'

VowelSignKettiPaaDashPilla

\u{dd4}: 'ු'

VowelSignDigaPaaDashPilla

\u{dd6}: 'ූ'

VowelSignGaettaDashPilla

\u{dd8}: 'ෘ'

VowelSignKombuva

\u{dd9}: 'ෙ'

VowelSignDigaKombuva

\u{dda}: 'ේ'

VowelSignKombuDeka

\u{ddb}: 'ෛ'

VowelSignKombuvaHaaAelaDashPilla

\u{ddc}: 'ො'

VowelSignKombuvaHaaDigaAelaDashPilla

\u{ddd}: 'ෝ'

VowelSignKombuvaHaaGayanukitta

\u{dde}: 'ෞ'

VowelSignGayanukitta

\u{ddf}: 'ෟ'

LithDigitZero

\u{de6}: '෦'

LithDigitOne

\u{de7}: '෧'

LithDigitTwo

\u{de8}: '෨'

LithDigitThree

\u{de9}: '෩'

LithDigitFour

\u{dea}: '෪'

LithDigitFive

\u{deb}: '෫'

LithDigitSix

\u{dec}: '෬'

LithDigitSeven

\u{ded}: '෭'

LithDigitEight

\u{dee}: '෮'

LithDigitNine

\u{def}: '෯'

VowelSignDigaGaettaDashPilla

\u{df2}: 'ෲ'

VowelSignDigaGayanukitta

\u{df3}: 'ෳ'

PunctuationKunddaliya

\u{df4}: '෴'

Methods

impl Sinhala

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case