Enum unicode_types::spacing_modifier_letters::SpacingModifierLetters

pub enum SpacingModifierLetters {
    ModifierLetterSmallH,
    ModifierLetterSmallHWithHook,
    ModifierLetterSmallJ,
    ModifierLetterSmallR,
    ModifierLetterSmallTurnedR,
    ModifierLetterSmallTurnedRWithHook,
    ModifierLetterSmallCapitalInvertedR,
    ModifierLetterSmallW,
    ModifierLetterSmallY,
    ModifierLetterPrime,
    ModifierLetterDoublePrime,
    ModifierLetterTurnedComma,
    ModifierLetterApostrophe,
    ModifierLetterReversedComma,
    ModifierLetterRightHalfRing,
    ModifierLetterLeftHalfRing,
    ModifierLetterGlottalStop,
    ModifierLetterReversedGlottalStop,
    ModifierLetterLeftArrowhead,
    ModifierLetterRightArrowhead,
    ModifierLetterUpArrowhead,
    ModifierLetterDownArrowhead,
    ModifierLetterCircumflexAccent,
    Caron,
    ModifierLetterVerticalLine,
    ModifierLetterMacron,
    ModifierLetterAcuteAccent,
    ModifierLetterGraveAccent,
    ModifierLetterLowVerticalLine,
    ModifierLetterLowMacron,
    ModifierLetterLowGraveAccent,
    ModifierLetterLowAcuteAccent,
    ModifierLetterTriangularColon,
    ModifierLetterHalfTriangularColon,
    ModifierLetterCentredRightHalfRing,
    ModifierLetterCentredLeftHalfRing,
    ModifierLetterUpTack,
    ModifierLetterDownTack,
    ModifierLetterPlusSign,
    ModifierLetterMinusSign,
    Breve,
    DotAbove,
    RingAbove,
    Ogonek,
    SmallTilde,
    DoubleAcuteAccent,
    ModifierLetterRhoticHook,
    ModifierLetterCrossAccent,
    ModifierLetterSmallGamma,
    ModifierLetterSmallL,
    ModifierLetterSmallS,
    ModifierLetterSmallX,
    ModifierLetterSmallReversedGlottalStop,
    ModifierLetterExtraDashHighToneBar,
    ModifierLetterHighToneBar,
    ModifierLetterMidToneBar,
    ModifierLetterLowToneBar,
    ModifierLetterExtraDashLowToneBar,
    ModifierLetterYinDepartingToneMark,
    ModifierLetterYangDepartingToneMark,
    ModifierLetterVoicing,
    ModifierLetterUnaspirated,
    ModifierLetterDoubleApostrophe,
    ModifierLetterLowDownArrowhead,
    ModifierLetterLowUpArrowhead,
    ModifierLetterLowLeftArrowhead,
    ModifierLetterLowRightArrowhead,
    ModifierLetterLowRing,
    ModifierLetterMiddleGraveAccent,
    ModifierLetterMiddleDoubleGraveAccent,
    ModifierLetterMiddleDoubleAcuteAccent,
    ModifierLetterLowTilde,
    ModifierLetterRaisedColon,
    ModifierLetterBeginHighTone,
    ModifierLetterEndHighTone,
    ModifierLetterBeginLowTone,
    ModifierLetterEndLowTone,
    ModifierLetterShelf,
    ModifierLetterOpenShelf,
}

An enum to represent all characters in the SpacingModifierLetters block.

Variants
ModifierLetterSmallH

\u{2b0}: 'ʰ'

ModifierLetterSmallHWithHook

\u{2b1}: 'ʱ'

ModifierLetterSmallJ

\u{2b2}: 'ʲ'

ModifierLetterSmallR

\u{2b3}: 'ʳ'

ModifierLetterSmallTurnedR

\u{2b4}: 'ʴ'

ModifierLetterSmallTurnedRWithHook

\u{2b5}: 'ʵ'

ModifierLetterSmallCapitalInvertedR

\u{2b6}: 'ʶ'

ModifierLetterSmallW

\u{2b7}: 'ʷ'

ModifierLetterSmallY

\u{2b8}: 'ʸ'

ModifierLetterPrime

\u{2b9}: 'ʹ'

ModifierLetterDoublePrime

\u{2ba}: 'ʺ'

ModifierLetterTurnedComma

\u{2bb}: 'ʻ'

ModifierLetterApostrophe

\u{2bc}: 'ʼ'

ModifierLetterReversedComma

\u{2bd}: 'ʽ'

ModifierLetterRightHalfRing

\u{2be}: 'ʾ'

ModifierLetterLeftHalfRing

\u{2bf}: 'ʿ'

ModifierLetterGlottalStop

\u{2c0}: 'ˀ'

ModifierLetterReversedGlottalStop

\u{2c1}: 'ˁ'

ModifierLetterLeftArrowhead

\u{2c2}: '˂'

ModifierLetterRightArrowhead

\u{2c3}: '˃'

ModifierLetterUpArrowhead

\u{2c4}: '˄'

ModifierLetterDownArrowhead

\u{2c5}: '˅'

ModifierLetterCircumflexAccent

\u{2c6}: 'ˆ'

Caron

\u{2c7}: 'ˇ'

ModifierLetterVerticalLine

\u{2c8}: 'ˈ'

ModifierLetterMacron

\u{2c9}: 'ˉ'

ModifierLetterAcuteAccent

\u{2ca}: 'ˊ'

ModifierLetterGraveAccent

\u{2cb}: 'ˋ'

ModifierLetterLowVerticalLine

\u{2cc}: 'ˌ'

ModifierLetterLowMacron

\u{2cd}: 'ˍ'

ModifierLetterLowGraveAccent

\u{2ce}: 'ˎ'

ModifierLetterLowAcuteAccent

\u{2cf}: 'ˏ'

ModifierLetterTriangularColon

\u{2d0}: 'ː'

ModifierLetterHalfTriangularColon

\u{2d1}: 'ˑ'

ModifierLetterCentredRightHalfRing

\u{2d2}: '˒'

ModifierLetterCentredLeftHalfRing

\u{2d3}: '˓'

ModifierLetterUpTack

\u{2d4}: '˔'

ModifierLetterDownTack

\u{2d5}: '˕'

ModifierLetterPlusSign

\u{2d6}: '˖'

ModifierLetterMinusSign

\u{2d7}: '˗'

Breve

\u{2d8}: '˘'

DotAbove

\u{2d9}: '˙'

RingAbove

\u{2da}: '˚'

Ogonek

\u{2db}: '˛'

SmallTilde

\u{2dc}: '˜'

DoubleAcuteAccent

\u{2dd}: '˝'

ModifierLetterRhoticHook

\u{2de}: '˞'

ModifierLetterCrossAccent

\u{2df}: '˟'

ModifierLetterSmallGamma

\u{2e0}: 'ˠ'

ModifierLetterSmallL

\u{2e1}: 'ˡ'

ModifierLetterSmallS

\u{2e2}: 'ˢ'

ModifierLetterSmallX

\u{2e3}: 'ˣ'

ModifierLetterSmallReversedGlottalStop

\u{2e4}: 'ˤ'

ModifierLetterExtraDashHighToneBar

\u{2e5}: '˥'

ModifierLetterHighToneBar

\u{2e6}: '˦'

ModifierLetterMidToneBar

\u{2e7}: '˧'

ModifierLetterLowToneBar

\u{2e8}: '˨'

ModifierLetterExtraDashLowToneBar

\u{2e9}: '˩'

ModifierLetterYinDepartingToneMark

\u{2ea}: '˪'

ModifierLetterYangDepartingToneMark

\u{2eb}: '˫'

ModifierLetterVoicing

\u{2ec}: 'ˬ'

ModifierLetterUnaspirated

\u{2ed}: '˭'

ModifierLetterDoubleApostrophe

\u{2ee}: 'ˮ'

ModifierLetterLowDownArrowhead

\u{2ef}: '˯'

ModifierLetterLowUpArrowhead

\u{2f0}: '˰'

ModifierLetterLowLeftArrowhead

\u{2f1}: '˱'

ModifierLetterLowRightArrowhead

\u{2f2}: '˲'

ModifierLetterLowRing

\u{2f3}: '˳'

ModifierLetterMiddleGraveAccent

\u{2f4}: '˴'

ModifierLetterMiddleDoubleGraveAccent

\u{2f5}: '˵'

ModifierLetterMiddleDoubleAcuteAccent

\u{2f6}: '˶'

ModifierLetterLowTilde

\u{2f7}: '˷'

ModifierLetterRaisedColon

\u{2f8}: '˸'

ModifierLetterBeginHighTone

\u{2f9}: '˹'

ModifierLetterEndHighTone

\u{2fa}: '˺'

ModifierLetterBeginLowTone

\u{2fb}: '˻'

ModifierLetterEndLowTone

\u{2fc}: '˼'

ModifierLetterShelf

\u{2fd}: '˽'

ModifierLetterOpenShelf

\u{2fe}: '˾'

Methods

impl SpacingModifierLetters

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case