Enum unicode_types::supplemental_arrows_b::SupplementalArrowsB

pub enum SupplementalArrowsB {
    RightwardsTwoDashHeadedArrowWithVerticalStroke,
    RightwardsTwoDashHeadedArrowWithDoubleVerticalStroke,
    LeftwardsDoubleArrowWithVerticalStroke,
    RightwardsDoubleArrowWithVerticalStroke,
    LeftRightDoubleArrowWithVerticalStroke,
    RightwardsTwoDashHeadedArrowFromBar,
    LeftwardsDoubleArrowFromBar,
    RightwardsDoubleArrowFromBar,
    DownwardsArrowWithHorizontalStroke,
    UpwardsArrowWithHorizontalStroke,
    UpwardsTripleArrow,
    DownwardsTripleArrow,
    LeftwardsDoubleDashArrow,
    RightwardsDoubleDashArrow,
    LeftwardsTripleDashArrow,
    RightwardsTripleDashArrow,
    RightwardsTwoDashHeadedTripleDashArrow,
    RightwardsArrowWithDottedStem,
    UpwardsArrowToBar,
    DownwardsArrowToBar,
    RightwardsArrowWithTailWithVerticalStroke,
    RightwardsArrowWithTailWithDoubleVerticalStroke,
    RightwardsTwoDashHeadedArrowWithTail,
    RightwardsTwoDashHeadedArrowWithTailWithVerticalStroke,
    RightwardsTwoDashHeadedArrowWithTailWithDoubleVerticalStroke,
    LeftwardsArrowDashTail,
    RightwardsArrowDashTail,
    LeftwardsDoubleArrowDashTail,
    RightwardsDoubleArrowDashTail,
    LeftwardsArrowToBlackDiamond,
    RightwardsArrowToBlackDiamond,
    LeftwardsArrowFromBarToBlackDiamond,
    RightwardsArrowFromBarToBlackDiamond,
    NorthWestAndSouthEastArrow,
    NorthEastAndSouthWestArrow,
    NorthWestArrowWithHook,
    NorthEastArrowWithHook,
    SouthEastArrowWithHook,
    SouthWestArrowWithHook,
    NorthWestArrowAndNorthEastArrow,
    NorthEastArrowAndSouthEastArrow,
    SouthEastArrowAndSouthWestArrow,
    SouthWestArrowAndNorthWestArrow,
    RisingDiagonalCrossingFallingDiagonal,
    FallingDiagonalCrossingRisingDiagonal,
    SouthEastArrowCrossingNorthEastArrow,
    NorthEastArrowCrossingSouthEastArrow,
    FallingDiagonalCrossingNorthEastArrow,
    RisingDiagonalCrossingSouthEastArrow,
    NorthEastArrowCrossingNorthWestArrow,
    NorthWestArrowCrossingNorthEastArrow,
    WaveArrowPointingDirectlyRight,
    ArrowPointingRightwardsThenCurvingUpwards,
    ArrowPointingRightwardsThenCurvingDownwards,
    ArrowPointingDownwardsThenCurvingLeftwards,
    ArrowPointingDownwardsThenCurvingRightwards,
    RightDashSideArcClockwiseArrow,
    LeftDashSideArcAnticlockwiseArrow,
    TopArcAnticlockwiseArrow,
    BottomArcAnticlockwiseArrow,
    TopArcClockwiseArrowWithMinus,
    TopArcAnticlockwiseArrowWithPlus,
    LowerRightSemicircularClockwiseArrow,
    LowerLeftSemicircularAnticlockwiseArrow,
    AnticlockwiseClosedCircleArrow,
    ClockwiseClosedCircleArrow,
    RightwardsArrowAboveShortLeftwardsArrow,
    LeftwardsArrowAboveShortRightwardsArrow,
    ShortRightwardsArrowAboveLeftwardsArrow,
    RightwardsArrowWithPlusBelow,
    LeftwardsArrowWithPlusBelow,
    RightwardsArrowThroughX,
    LeftRightArrowThroughSmallCircle,
    UpwardsTwoDashHeadedArrowFromSmallCircle,
    LeftBarbUpRightBarbDownHarpoon,
    LeftBarbDownRightBarbUpHarpoon,
    UpBarbRightDownBarbLeftHarpoon,
    UpBarbLeftDownBarbRightHarpoon,
    LeftBarbUpRightBarbUpHarpoon,
    UpBarbRightDownBarbRightHarpoon,
    LeftBarbDownRightBarbDownHarpoon,
    UpBarbLeftDownBarbLeftHarpoon,
    LeftwardsHarpoonWithBarbUpToBar,
    RightwardsHarpoonWithBarbUpToBar,
    UpwardsHarpoonWithBarbRightToBar,
    DownwardsHarpoonWithBarbRightToBar,
    LeftwardsHarpoonWithBarbDownToBar,
    RightwardsHarpoonWithBarbDownToBar,
    UpwardsHarpoonWithBarbLeftToBar,
    DownwardsHarpoonWithBarbLeftToBar,
    LeftwardsHarpoonWithBarbUpFromBar,
    RightwardsHarpoonWithBarbUpFromBar,
    UpwardsHarpoonWithBarbRightFromBar,
    DownwardsHarpoonWithBarbRightFromBar,
    LeftwardsHarpoonWithBarbDownFromBar,
    RightwardsHarpoonWithBarbDownFromBar,
    UpwardsHarpoonWithBarbLeftFromBar,
    DownwardsHarpoonWithBarbLeftFromBar,
    LeftwardsHarpoonWithBarbUpAboveLeftwardsHarpoonWithBarbDown,
    UpwardsHarpoonWithBarbLeftBesideUpwardsHarpoonWithBarbRight,
    RightwardsHarpoonWithBarbUpAboveRightwardsHarpoonWithBarbDown,
    DownwardsHarpoonWithBarbLeftBesideDownwardsHarpoonWithBarbRight,
    LeftwardsHarpoonWithBarbUpAboveRightwardsHarpoonWithBarbUp,
    LeftwardsHarpoonWithBarbDownAboveRightwardsHarpoonWithBarbDown,
    RightwardsHarpoonWithBarbUpAboveLeftwardsHarpoonWithBarbUp,
    RightwardsHarpoonWithBarbDownAboveLeftwardsHarpoonWithBarbDown,
    LeftwardsHarpoonWithBarbUpAboveLongDash,
    LeftwardsHarpoonWithBarbDownBelowLongDash,
    RightwardsHarpoonWithBarbUpAboveLongDash,
    RightwardsHarpoonWithBarbDownBelowLongDash,
    UpwardsHarpoonWithBarbLeftBesideDownwardsHarpoonWithBarbRight,
    DownwardsHarpoonWithBarbLeftBesideUpwardsHarpoonWithBarbRight,
    RightDoubleArrowWithRoundedHead,
    EqualsSignAboveRightwardsArrow,
    TildeOperatorAboveRightwardsArrow,
    LeftwardsArrowAboveTildeOperator,
    RightwardsArrowAboveTildeOperator,
    RightwardsArrowAboveAlmostEqualTo,
    LessDashThanAboveLeftwardsArrow,
    LeftwardsArrowThroughLessDashThan,
    GreaterDashThanAboveRightwardsArrow,
    SubsetAboveRightwardsArrow,
    LeftwardsArrowThroughSubset,
    SupersetAboveLeftwardsArrow,
    LeftFishTail,
    RightFishTail,
    UpFishTail,
}

An enum to represent all characters in the SupplementalArrowsB block.

Variants
RightwardsTwoDashHeadedArrowWithVerticalStroke

\u{2900}: '⤀'

RightwardsTwoDashHeadedArrowWithDoubleVerticalStroke

\u{2901}: '⤁'

LeftwardsDoubleArrowWithVerticalStroke

\u{2902}: '⤂'

RightwardsDoubleArrowWithVerticalStroke

\u{2903}: '⤃'

LeftRightDoubleArrowWithVerticalStroke

\u{2904}: '⤄'

RightwardsTwoDashHeadedArrowFromBar

\u{2905}: '⤅'

LeftwardsDoubleArrowFromBar

\u{2906}: '⤆'

RightwardsDoubleArrowFromBar

\u{2907}: '⤇'

DownwardsArrowWithHorizontalStroke

\u{2908}: '⤈'

UpwardsArrowWithHorizontalStroke

\u{2909}: '⤉'

UpwardsTripleArrow

\u{290a}: '⤊'

DownwardsTripleArrow

\u{290b}: '⤋'

LeftwardsDoubleDashArrow

\u{290c}: '⤌'

RightwardsDoubleDashArrow

\u{290d}: '⤍'

LeftwardsTripleDashArrow

\u{290e}: '⤎'

RightwardsTripleDashArrow

\u{290f}: '⤏'

RightwardsTwoDashHeadedTripleDashArrow

\u{2910}: '⤐'

RightwardsArrowWithDottedStem

\u{2911}: '⤑'

UpwardsArrowToBar

\u{2912}: '⤒'

DownwardsArrowToBar

\u{2913}: '⤓'

RightwardsArrowWithTailWithVerticalStroke

\u{2914}: '⤔'

RightwardsArrowWithTailWithDoubleVerticalStroke

\u{2915}: '⤕'

RightwardsTwoDashHeadedArrowWithTail

\u{2916}: '⤖'

RightwardsTwoDashHeadedArrowWithTailWithVerticalStroke

\u{2917}: '⤗'

RightwardsTwoDashHeadedArrowWithTailWithDoubleVerticalStroke

\u{2918}: '⤘'

LeftwardsArrowDashTail

\u{2919}: '⤙'

RightwardsArrowDashTail

\u{291a}: '⤚'

LeftwardsDoubleArrowDashTail

\u{291b}: '⤛'

RightwardsDoubleArrowDashTail

\u{291c}: '⤜'

LeftwardsArrowToBlackDiamond

\u{291d}: '⤝'

RightwardsArrowToBlackDiamond

\u{291e}: '⤞'

LeftwardsArrowFromBarToBlackDiamond

\u{291f}: '⤟'

RightwardsArrowFromBarToBlackDiamond

\u{2920}: '⤠'

NorthWestAndSouthEastArrow

\u{2921}: '⤡'

NorthEastAndSouthWestArrow

\u{2922}: '⤢'

NorthWestArrowWithHook

\u{2923}: '⤣'

NorthEastArrowWithHook

\u{2924}: '⤤'

SouthEastArrowWithHook

\u{2925}: '⤥'

SouthWestArrowWithHook

\u{2926}: '⤦'

NorthWestArrowAndNorthEastArrow

\u{2927}: '⤧'

NorthEastArrowAndSouthEastArrow

\u{2928}: '⤨'

SouthEastArrowAndSouthWestArrow

\u{2929}: '⤩'

SouthWestArrowAndNorthWestArrow

\u{292a}: '⤪'

RisingDiagonalCrossingFallingDiagonal

\u{292b}: '⤫'

FallingDiagonalCrossingRisingDiagonal

\u{292c}: '⤬'

SouthEastArrowCrossingNorthEastArrow

\u{292d}: '⤭'

NorthEastArrowCrossingSouthEastArrow

\u{292e}: '⤮'

FallingDiagonalCrossingNorthEastArrow

\u{292f}: '⤯'

RisingDiagonalCrossingSouthEastArrow

\u{2930}: '⤰'

NorthEastArrowCrossingNorthWestArrow

\u{2931}: '⤱'

NorthWestArrowCrossingNorthEastArrow

\u{2932}: '⤲'

WaveArrowPointingDirectlyRight

\u{2933}: '⤳'

ArrowPointingRightwardsThenCurvingUpwards

\u{2934}: '⤴'

ArrowPointingRightwardsThenCurvingDownwards

\u{2935}: '⤵'

ArrowPointingDownwardsThenCurvingLeftwards

\u{2936}: '⤶'

ArrowPointingDownwardsThenCurvingRightwards

\u{2937}: '⤷'

RightDashSideArcClockwiseArrow

\u{2938}: '⤸'

LeftDashSideArcAnticlockwiseArrow

\u{2939}: '⤹'

TopArcAnticlockwiseArrow

\u{293a}: '⤺'

BottomArcAnticlockwiseArrow

\u{293b}: '⤻'

TopArcClockwiseArrowWithMinus

\u{293c}: '⤼'

TopArcAnticlockwiseArrowWithPlus

\u{293d}: '⤽'

LowerRightSemicircularClockwiseArrow

\u{293e}: '⤾'

LowerLeftSemicircularAnticlockwiseArrow

\u{293f}: '⤿'

AnticlockwiseClosedCircleArrow

\u{2940}: '⥀'

ClockwiseClosedCircleArrow

\u{2941}: '⥁'

RightwardsArrowAboveShortLeftwardsArrow

\u{2942}: '⥂'

LeftwardsArrowAboveShortRightwardsArrow

\u{2943}: '⥃'

ShortRightwardsArrowAboveLeftwardsArrow

\u{2944}: '⥄'

RightwardsArrowWithPlusBelow

\u{2945}: '⥅'

LeftwardsArrowWithPlusBelow

\u{2946}: '⥆'

RightwardsArrowThroughX

\u{2947}: '⥇'

LeftRightArrowThroughSmallCircle

\u{2948}: '⥈'

UpwardsTwoDashHeadedArrowFromSmallCircle

\u{2949}: '⥉'

LeftBarbUpRightBarbDownHarpoon

\u{294a}: '⥊'

LeftBarbDownRightBarbUpHarpoon

\u{294b}: '⥋'

UpBarbRightDownBarbLeftHarpoon

\u{294c}: '⥌'

UpBarbLeftDownBarbRightHarpoon

\u{294d}: '⥍'

LeftBarbUpRightBarbUpHarpoon

\u{294e}: '⥎'

UpBarbRightDownBarbRightHarpoon

\u{294f}: '⥏'

LeftBarbDownRightBarbDownHarpoon

\u{2950}: '⥐'

UpBarbLeftDownBarbLeftHarpoon

\u{2951}: '⥑'

LeftwardsHarpoonWithBarbUpToBar

\u{2952}: '⥒'

RightwardsHarpoonWithBarbUpToBar

\u{2953}: '⥓'

UpwardsHarpoonWithBarbRightToBar

\u{2954}: '⥔'

DownwardsHarpoonWithBarbRightToBar

\u{2955}: '⥕'

LeftwardsHarpoonWithBarbDownToBar

\u{2956}: '⥖'

RightwardsHarpoonWithBarbDownToBar

\u{2957}: '⥗'

UpwardsHarpoonWithBarbLeftToBar

\u{2958}: '⥘'

DownwardsHarpoonWithBarbLeftToBar

\u{2959}: '⥙'

LeftwardsHarpoonWithBarbUpFromBar

\u{295a}: '⥚'

RightwardsHarpoonWithBarbUpFromBar

\u{295b}: '⥛'

UpwardsHarpoonWithBarbRightFromBar

\u{295c}: '⥜'

DownwardsHarpoonWithBarbRightFromBar

\u{295d}: '⥝'

LeftwardsHarpoonWithBarbDownFromBar

\u{295e}: '⥞'

RightwardsHarpoonWithBarbDownFromBar

\u{295f}: '⥟'

UpwardsHarpoonWithBarbLeftFromBar

\u{2960}: '⥠'

DownwardsHarpoonWithBarbLeftFromBar

\u{2961}: '⥡'

LeftwardsHarpoonWithBarbUpAboveLeftwardsHarpoonWithBarbDown

\u{2962}: '⥢'

UpwardsHarpoonWithBarbLeftBesideUpwardsHarpoonWithBarbRight

\u{2963}: '⥣'

RightwardsHarpoonWithBarbUpAboveRightwardsHarpoonWithBarbDown

\u{2964}: '⥤'

DownwardsHarpoonWithBarbLeftBesideDownwardsHarpoonWithBarbRight

\u{2965}: '⥥'

LeftwardsHarpoonWithBarbUpAboveRightwardsHarpoonWithBarbUp

\u{2966}: '⥦'

LeftwardsHarpoonWithBarbDownAboveRightwardsHarpoonWithBarbDown

\u{2967}: '⥧'

RightwardsHarpoonWithBarbUpAboveLeftwardsHarpoonWithBarbUp

\u{2968}: '⥨'

RightwardsHarpoonWithBarbDownAboveLeftwardsHarpoonWithBarbDown

\u{2969}: '⥩'

LeftwardsHarpoonWithBarbUpAboveLongDash

\u{296a}: '⥪'

LeftwardsHarpoonWithBarbDownBelowLongDash

\u{296b}: '⥫'

RightwardsHarpoonWithBarbUpAboveLongDash

\u{296c}: '⥬'

RightwardsHarpoonWithBarbDownBelowLongDash

\u{296d}: '⥭'

UpwardsHarpoonWithBarbLeftBesideDownwardsHarpoonWithBarbRight

\u{296e}: '⥮'

DownwardsHarpoonWithBarbLeftBesideUpwardsHarpoonWithBarbRight

\u{296f}: '⥯'

RightDoubleArrowWithRoundedHead

\u{2970}: '⥰'

EqualsSignAboveRightwardsArrow

\u{2971}: '⥱'

TildeOperatorAboveRightwardsArrow

\u{2972}: '⥲'

LeftwardsArrowAboveTildeOperator

\u{2973}: '⥳'

RightwardsArrowAboveTildeOperator

\u{2974}: '⥴'

RightwardsArrowAboveAlmostEqualTo

\u{2975}: '⥵'

LessDashThanAboveLeftwardsArrow

\u{2976}: '⥶'

LeftwardsArrowThroughLessDashThan

\u{2977}: '⥷'

GreaterDashThanAboveRightwardsArrow

\u{2978}: '⥸'

SubsetAboveRightwardsArrow

\u{2979}: '⥹'

LeftwardsArrowThroughSubset

\u{297a}: '⥺'

SupersetAboveLeftwardsArrow

\u{297b}: '⥻'

LeftFishTail

\u{297c}: '⥼'

RightFishTail

\u{297d}: '⥽'

UpFishTail

\u{297e}: '⥾'

Methods

impl SupplementalArrowsB

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case