Enum unicode_types::supplemental_arrows_c::SupplementalArrowsC

pub enum SupplementalArrowsC {
    LeftwardsArrowWithSmallTriangleArrowhead,
    UpwardsArrowWithSmallTriangleArrowhead,
    RightwardsArrowWithSmallTriangleArrowhead,
    DownwardsArrowWithSmallTriangleArrowhead,
    LeftwardsArrowWithMediumTriangleArrowhead,
    UpwardsArrowWithMediumTriangleArrowhead,
    RightwardsArrowWithMediumTriangleArrowhead,
    DownwardsArrowWithMediumTriangleArrowhead,
    LeftwardsArrowWithLargeTriangleArrowhead,
    UpwardsArrowWithLargeTriangleArrowhead,
    RightwardsArrowWithLargeTriangleArrowhead,
    DownwardsArrowWithLargeTriangleArrowhead,
    LeftwardsArrowWithSmallEquilateralArrowhead,
    UpwardsArrowWithSmallEquilateralArrowhead,
    RightwardsArrowWithSmallEquilateralArrowhead,
    DownwardsArrowWithSmallEquilateralArrowhead,
    LeftwardsArrowWithEquilateralArrowhead,
    UpwardsArrowWithEquilateralArrowhead,
    RightwardsArrowWithEquilateralArrowhead,
    DownwardsArrowWithEquilateralArrowhead,
    HeavyLeftwardsArrowWithEquilateralArrowhead,
    HeavyUpwardsArrowWithEquilateralArrowhead,
    HeavyRightwardsArrowWithEquilateralArrowhead,
    HeavyDownwardsArrowWithEquilateralArrowhead,
    HeavyLeftwardsArrowWithLargeEquilateralArrowhead,
    HeavyUpwardsArrowWithLargeEquilateralArrowhead,
    HeavyRightwardsArrowWithLargeEquilateralArrowhead,
    HeavyDownwardsArrowWithLargeEquilateralArrowhead,
    LeftwardsTriangleDashHeadedArrowWithNarrowShaft,
    UpwardsTriangleDashHeadedArrowWithNarrowShaft,
    RightwardsTriangleDashHeadedArrowWithNarrowShaft,
    DownwardsTriangleDashHeadedArrowWithNarrowShaft,
    LeftwardsTriangleDashHeadedArrowWithMediumShaft,
    UpwardsTriangleDashHeadedArrowWithMediumShaft,
    RightwardsTriangleDashHeadedArrowWithMediumShaft,
    DownwardsTriangleDashHeadedArrowWithMediumShaft,
    LeftwardsTriangleDashHeadedArrowWithBoldShaft,
    UpwardsTriangleDashHeadedArrowWithBoldShaft,
    RightwardsTriangleDashHeadedArrowWithBoldShaft,
    DownwardsTriangleDashHeadedArrowWithBoldShaft,
    LeftwardsTriangleDashHeadedArrowWithHeavyShaft,
    UpwardsTriangleDashHeadedArrowWithHeavyShaft,
    RightwardsTriangleDashHeadedArrowWithHeavyShaft,
    DownwardsTriangleDashHeadedArrowWithHeavyShaft,
    LeftwardsTriangleDashHeadedArrowWithVeryHeavyShaft,
    UpwardsTriangleDashHeadedArrowWithVeryHeavyShaft,
    RightwardsTriangleDashHeadedArrowWithVeryHeavyShaft,
    DownwardsTriangleDashHeadedArrowWithVeryHeavyShaft,
    LeftwardsFingerDashPostArrow,
    UpwardsFingerDashPostArrow,
    RightwardsFingerDashPostArrow,
    DownwardsFingerDashPostArrow,
    LeftwardsSquaredArrow,
    UpwardsSquaredArrow,
    RightwardsSquaredArrow,
    DownwardsSquaredArrow,
    LeftwardsCompressedArrow,
    UpwardsCompressedArrow,
    RightwardsCompressedArrow,
    DownwardsCompressedArrow,
    LeftwardsHeavyCompressedArrow,
    UpwardsHeavyCompressedArrow,
    RightwardsHeavyCompressedArrow,
    DownwardsHeavyCompressedArrow,
    LeftwardsHeavyArrow,
    UpwardsHeavyArrow,
    RightwardsHeavyArrow,
    DownwardsHeavyArrow,
    LeftwardsSansDashSerifArrow,
    UpwardsSansDashSerifArrow,
    RightwardsSansDashSerifArrow,
    DownwardsSansDashSerifArrow,
    NorthWestSansDashSerifArrow,
    NorthEastSansDashSerifArrow,
    SouthEastSansDashSerifArrow,
    SouthWestSansDashSerifArrow,
    LeftRightSansDashSerifArrow,
    UpDownSansDashSerifArrow,
    WideDashHeadedLeftwardsLightBarbArrow,
    WideDashHeadedUpwardsLightBarbArrow,
    WideDashHeadedRightwardsLightBarbArrow,
    WideDashHeadedDownwardsLightBarbArrow,
    WideDashHeadedNorthWestLightBarbArrow,
    WideDashHeadedNorthEastLightBarbArrow,
    WideDashHeadedSouthEastLightBarbArrow,
    WideDashHeadedSouthWestLightBarbArrow,
    WideDashHeadedLeftwardsBarbArrow,
    WideDashHeadedUpwardsBarbArrow,
    WideDashHeadedRightwardsBarbArrow,
    WideDashHeadedDownwardsBarbArrow,
    WideDashHeadedNorthWestBarbArrow,
    WideDashHeadedNorthEastBarbArrow,
    WideDashHeadedSouthEastBarbArrow,
    WideDashHeadedSouthWestBarbArrow,
    WideDashHeadedLeftwardsMediumBarbArrow,
    WideDashHeadedUpwardsMediumBarbArrow,
    WideDashHeadedRightwardsMediumBarbArrow,
    WideDashHeadedDownwardsMediumBarbArrow,
    WideDashHeadedNorthWestMediumBarbArrow,
    WideDashHeadedNorthEastMediumBarbArrow,
    WideDashHeadedSouthEastMediumBarbArrow,
    WideDashHeadedSouthWestMediumBarbArrow,
    WideDashHeadedLeftwardsHeavyBarbArrow,
    WideDashHeadedUpwardsHeavyBarbArrow,
    WideDashHeadedRightwardsHeavyBarbArrow,
    WideDashHeadedDownwardsHeavyBarbArrow,
    WideDashHeadedNorthWestHeavyBarbArrow,
    WideDashHeadedNorthEastHeavyBarbArrow,
    WideDashHeadedSouthEastHeavyBarbArrow,
    WideDashHeadedSouthWestHeavyBarbArrow,
    WideDashHeadedLeftwardsVeryHeavyBarbArrow,
    WideDashHeadedUpwardsVeryHeavyBarbArrow,
    WideDashHeadedRightwardsVeryHeavyBarbArrow,
    WideDashHeadedDownwardsVeryHeavyBarbArrow,
    WideDashHeadedNorthWestVeryHeavyBarbArrow,
    WideDashHeadedNorthEastVeryHeavyBarbArrow,
    WideDashHeadedSouthEastVeryHeavyBarbArrow,
    WideDashHeadedSouthWestVeryHeavyBarbArrow,
    LeftwardsTriangleArrowhead,
    UpwardsTriangleArrowhead,
    RightwardsTriangleArrowhead,
    DownwardsTriangleArrowhead,
    LeftwardsWhiteArrowWithinTriangleArrowhead,
    UpwardsWhiteArrowWithinTriangleArrowhead,
    RightwardsWhiteArrowWithinTriangleArrowhead,
    DownwardsWhiteArrowWithinTriangleArrowhead,
    LeftwardsArrowWithNotchedTail,
    UpwardsArrowWithNotchedTail,
    RightwardsArrowWithNotchedTail,
    DownwardsArrowWithNotchedTail,
    HeavyArrowShaftWidthOne,
    HeavyArrowShaftWidthTwoThirds,
    HeavyArrowShaftWidthOneHalf,
    HeavyArrowShaftWidthOneThird,
    LeftwardsBottomDashShadedWhiteArrow,
    RightwardsBottomShadedWhiteArrow,
    LeftwardsTopShadedWhiteArrow,
    RightwardsTopShadedWhiteArrow,
    LeftwardsLeftDashShadedWhiteArrow,
    RightwardsRightDashShadedWhiteArrow,
    LeftwardsRightDashShadedWhiteArrow,
    RightwardsLeftDashShadedWhiteArrow,
    LeftwardsBackDashTiltedShadowedWhiteArrow,
    RightwardsBackDashTiltedShadowedWhiteArrow,
    LeftwardsFrontDashTiltedShadowedWhiteArrow,
    RightwardsFrontDashTiltedShadowedWhiteArrow,
    WhiteArrowShaftWidthOne,
    WhiteArrowShaftWidthTwoThirds,
}

An enum to represent all characters in the SupplementalArrowsC block.

Variants
LeftwardsArrowWithSmallTriangleArrowhead

\u{1f800}: '🠀'

UpwardsArrowWithSmallTriangleArrowhead

\u{1f801}: '🠁'

RightwardsArrowWithSmallTriangleArrowhead

\u{1f802}: '🠂'

DownwardsArrowWithSmallTriangleArrowhead

\u{1f803}: '🠃'

LeftwardsArrowWithMediumTriangleArrowhead

\u{1f804}: '🠄'

UpwardsArrowWithMediumTriangleArrowhead

\u{1f805}: '🠅'

RightwardsArrowWithMediumTriangleArrowhead

\u{1f806}: '🠆'

DownwardsArrowWithMediumTriangleArrowhead

\u{1f807}: '🠇'

LeftwardsArrowWithLargeTriangleArrowhead

\u{1f808}: '🠈'

UpwardsArrowWithLargeTriangleArrowhead

\u{1f809}: '🠉'

RightwardsArrowWithLargeTriangleArrowhead

\u{1f80a}: '🠊'

DownwardsArrowWithLargeTriangleArrowhead

\u{1f80b}: '🠋'

LeftwardsArrowWithSmallEquilateralArrowhead

\u{1f810}: '🠐'

UpwardsArrowWithSmallEquilateralArrowhead

\u{1f811}: '🠑'

RightwardsArrowWithSmallEquilateralArrowhead

\u{1f812}: '🠒'

DownwardsArrowWithSmallEquilateralArrowhead

\u{1f813}: '🠓'

LeftwardsArrowWithEquilateralArrowhead

\u{1f814}: '🠔'

UpwardsArrowWithEquilateralArrowhead

\u{1f815}: '🠕'

RightwardsArrowWithEquilateralArrowhead

\u{1f816}: '🠖'

DownwardsArrowWithEquilateralArrowhead

\u{1f817}: '🠗'

HeavyLeftwardsArrowWithEquilateralArrowhead

\u{1f818}: '🠘'

HeavyUpwardsArrowWithEquilateralArrowhead

\u{1f819}: '🠙'

HeavyRightwardsArrowWithEquilateralArrowhead

\u{1f81a}: '🠚'

HeavyDownwardsArrowWithEquilateralArrowhead

\u{1f81b}: '🠛'

HeavyLeftwardsArrowWithLargeEquilateralArrowhead

\u{1f81c}: '🠜'

HeavyUpwardsArrowWithLargeEquilateralArrowhead

\u{1f81d}: '🠝'

HeavyRightwardsArrowWithLargeEquilateralArrowhead

\u{1f81e}: '🠞'

HeavyDownwardsArrowWithLargeEquilateralArrowhead

\u{1f81f}: '🠟'

LeftwardsTriangleDashHeadedArrowWithNarrowShaft

\u{1f820}: '🠠'

UpwardsTriangleDashHeadedArrowWithNarrowShaft

\u{1f821}: '🠡'

RightwardsTriangleDashHeadedArrowWithNarrowShaft

\u{1f822}: '🠢'

DownwardsTriangleDashHeadedArrowWithNarrowShaft

\u{1f823}: '🠣'

LeftwardsTriangleDashHeadedArrowWithMediumShaft

\u{1f824}: '🠤'

UpwardsTriangleDashHeadedArrowWithMediumShaft

\u{1f825}: '🠥'

RightwardsTriangleDashHeadedArrowWithMediumShaft

\u{1f826}: '🠦'

DownwardsTriangleDashHeadedArrowWithMediumShaft

\u{1f827}: '🠧'

LeftwardsTriangleDashHeadedArrowWithBoldShaft

\u{1f828}: '🠨'

UpwardsTriangleDashHeadedArrowWithBoldShaft

\u{1f829}: '🠩'

RightwardsTriangleDashHeadedArrowWithBoldShaft

\u{1f82a}: '🠪'

DownwardsTriangleDashHeadedArrowWithBoldShaft

\u{1f82b}: '🠫'

LeftwardsTriangleDashHeadedArrowWithHeavyShaft

\u{1f82c}: '🠬'

UpwardsTriangleDashHeadedArrowWithHeavyShaft

\u{1f82d}: '🠭'

RightwardsTriangleDashHeadedArrowWithHeavyShaft

\u{1f82e}: '🠮'

DownwardsTriangleDashHeadedArrowWithHeavyShaft

\u{1f82f}: '🠯'

LeftwardsTriangleDashHeadedArrowWithVeryHeavyShaft

\u{1f830}: '🠰'

UpwardsTriangleDashHeadedArrowWithVeryHeavyShaft

\u{1f831}: '🠱'

RightwardsTriangleDashHeadedArrowWithVeryHeavyShaft

\u{1f832}: '🠲'

DownwardsTriangleDashHeadedArrowWithVeryHeavyShaft

\u{1f833}: '🠳'

LeftwardsFingerDashPostArrow

\u{1f834}: '🠴'

UpwardsFingerDashPostArrow

\u{1f835}: '🠵'

RightwardsFingerDashPostArrow

\u{1f836}: '🠶'

DownwardsFingerDashPostArrow

\u{1f837}: '🠷'

LeftwardsSquaredArrow

\u{1f838}: '🠸'

UpwardsSquaredArrow

\u{1f839}: '🠹'

RightwardsSquaredArrow

\u{1f83a}: '🠺'

DownwardsSquaredArrow

\u{1f83b}: '🠻'

LeftwardsCompressedArrow

\u{1f83c}: '🠼'

UpwardsCompressedArrow

\u{1f83d}: '🠽'

RightwardsCompressedArrow

\u{1f83e}: '🠾'

DownwardsCompressedArrow

\u{1f83f}: '🠿'

LeftwardsHeavyCompressedArrow

\u{1f840}: '🡀'

UpwardsHeavyCompressedArrow

\u{1f841}: '🡁'

RightwardsHeavyCompressedArrow

\u{1f842}: '🡂'

DownwardsHeavyCompressedArrow

\u{1f843}: '🡃'

LeftwardsHeavyArrow

\u{1f844}: '🡄'

UpwardsHeavyArrow

\u{1f845}: '🡅'

RightwardsHeavyArrow

\u{1f846}: '🡆'

DownwardsHeavyArrow

\u{1f847}: '🡇'

LeftwardsSansDashSerifArrow

\u{1f850}: '🡐'

UpwardsSansDashSerifArrow

\u{1f851}: '🡑'

RightwardsSansDashSerifArrow

\u{1f852}: '🡒'

DownwardsSansDashSerifArrow

\u{1f853}: '🡓'

NorthWestSansDashSerifArrow

\u{1f854}: '🡔'

NorthEastSansDashSerifArrow

\u{1f855}: '🡕'

SouthEastSansDashSerifArrow

\u{1f856}: '🡖'

SouthWestSansDashSerifArrow

\u{1f857}: '🡗'

LeftRightSansDashSerifArrow

\u{1f858}: '🡘'

UpDownSansDashSerifArrow

\u{1f859}: '🡙'

WideDashHeadedLeftwardsLightBarbArrow

\u{1f860}: '🡠'

WideDashHeadedUpwardsLightBarbArrow

\u{1f861}: '🡡'

WideDashHeadedRightwardsLightBarbArrow

\u{1f862}: '🡢'

WideDashHeadedDownwardsLightBarbArrow

\u{1f863}: '🡣'

WideDashHeadedNorthWestLightBarbArrow

\u{1f864}: '🡤'

WideDashHeadedNorthEastLightBarbArrow

\u{1f865}: '🡥'

WideDashHeadedSouthEastLightBarbArrow

\u{1f866}: '🡦'

WideDashHeadedSouthWestLightBarbArrow

\u{1f867}: '🡧'

WideDashHeadedLeftwardsBarbArrow

\u{1f868}: '🡨'

WideDashHeadedUpwardsBarbArrow

\u{1f869}: '🡩'

WideDashHeadedRightwardsBarbArrow

\u{1f86a}: '🡪'

WideDashHeadedDownwardsBarbArrow

\u{1f86b}: '🡫'

WideDashHeadedNorthWestBarbArrow

\u{1f86c}: '🡬'

WideDashHeadedNorthEastBarbArrow

\u{1f86d}: '🡭'

WideDashHeadedSouthEastBarbArrow

\u{1f86e}: '🡮'

WideDashHeadedSouthWestBarbArrow

\u{1f86f}: '🡯'

WideDashHeadedLeftwardsMediumBarbArrow

\u{1f870}: '🡰'

WideDashHeadedUpwardsMediumBarbArrow

\u{1f871}: '🡱'

WideDashHeadedRightwardsMediumBarbArrow

\u{1f872}: '🡲'

WideDashHeadedDownwardsMediumBarbArrow

\u{1f873}: '🡳'

WideDashHeadedNorthWestMediumBarbArrow

\u{1f874}: '🡴'

WideDashHeadedNorthEastMediumBarbArrow

\u{1f875}: '🡵'

WideDashHeadedSouthEastMediumBarbArrow

\u{1f876}: '🡶'

WideDashHeadedSouthWestMediumBarbArrow

\u{1f877}: '🡷'

WideDashHeadedLeftwardsHeavyBarbArrow

\u{1f878}: '🡸'

WideDashHeadedUpwardsHeavyBarbArrow

\u{1f879}: '🡹'

WideDashHeadedRightwardsHeavyBarbArrow

\u{1f87a}: '🡺'

WideDashHeadedDownwardsHeavyBarbArrow

\u{1f87b}: '🡻'

WideDashHeadedNorthWestHeavyBarbArrow

\u{1f87c}: '🡼'

WideDashHeadedNorthEastHeavyBarbArrow

\u{1f87d}: '🡽'

WideDashHeadedSouthEastHeavyBarbArrow

\u{1f87e}: '🡾'

WideDashHeadedSouthWestHeavyBarbArrow

\u{1f87f}: '🡿'

WideDashHeadedLeftwardsVeryHeavyBarbArrow

\u{1f880}: '🢀'

WideDashHeadedUpwardsVeryHeavyBarbArrow

\u{1f881}: '🢁'

WideDashHeadedRightwardsVeryHeavyBarbArrow

\u{1f882}: '🢂'

WideDashHeadedDownwardsVeryHeavyBarbArrow

\u{1f883}: '🢃'

WideDashHeadedNorthWestVeryHeavyBarbArrow

\u{1f884}: '🢄'

WideDashHeadedNorthEastVeryHeavyBarbArrow

\u{1f885}: '🢅'

WideDashHeadedSouthEastVeryHeavyBarbArrow

\u{1f886}: '🢆'

WideDashHeadedSouthWestVeryHeavyBarbArrow

\u{1f887}: '🢇'

LeftwardsTriangleArrowhead

\u{1f890}: '🢐'

UpwardsTriangleArrowhead

\u{1f891}: '🢑'

RightwardsTriangleArrowhead

\u{1f892}: '🢒'

DownwardsTriangleArrowhead

\u{1f893}: '🢓'

LeftwardsWhiteArrowWithinTriangleArrowhead

\u{1f894}: '🢔'

UpwardsWhiteArrowWithinTriangleArrowhead

\u{1f895}: '🢕'

RightwardsWhiteArrowWithinTriangleArrowhead

\u{1f896}: '🢖'

DownwardsWhiteArrowWithinTriangleArrowhead

\u{1f897}: '🢗'

LeftwardsArrowWithNotchedTail

\u{1f898}: '🢘'

UpwardsArrowWithNotchedTail

\u{1f899}: '🢙'

RightwardsArrowWithNotchedTail

\u{1f89a}: '🢚'

DownwardsArrowWithNotchedTail

\u{1f89b}: '🢛'

HeavyArrowShaftWidthOne

\u{1f89c}: '🢜'

HeavyArrowShaftWidthTwoThirds

\u{1f89d}: '🢝'

HeavyArrowShaftWidthOneHalf

\u{1f89e}: '🢞'

HeavyArrowShaftWidthOneThird

\u{1f89f}: '🢟'

LeftwardsBottomDashShadedWhiteArrow

\u{1f8a0}: '🢠'

RightwardsBottomShadedWhiteArrow

\u{1f8a1}: '🢡'

LeftwardsTopShadedWhiteArrow

\u{1f8a2}: '🢢'

RightwardsTopShadedWhiteArrow

\u{1f8a3}: '🢣'

LeftwardsLeftDashShadedWhiteArrow

\u{1f8a4}: '🢤'

RightwardsRightDashShadedWhiteArrow

\u{1f8a5}: '🢥'

LeftwardsRightDashShadedWhiteArrow

\u{1f8a6}: '🢦'

RightwardsLeftDashShadedWhiteArrow

\u{1f8a7}: '🢧'

LeftwardsBackDashTiltedShadowedWhiteArrow

\u{1f8a8}: '🢨'

RightwardsBackDashTiltedShadowedWhiteArrow

\u{1f8a9}: '🢩'

LeftwardsFrontDashTiltedShadowedWhiteArrow

\u{1f8aa}: '🢪'

RightwardsFrontDashTiltedShadowedWhiteArrow

\u{1f8ab}: '🢫'

WhiteArrowShaftWidthOne

\u{1f8ac}: '🢬'

WhiteArrowShaftWidthTwoThirds

\u{1f8ad}: '🢭'

Methods

impl SupplementalArrowsC

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case