Enum unicode_types::supplemental_mathematical_operators::SupplementalMathematicalOperators

pub enum SupplementalMathematicalOperators {
    NDashAryCircledDotOperator,
    NDashAryCircledPlusOperator,
    NDashAryCircledTimesOperator,
    NDashAryUnionOperatorWithDot,
    NDashAryUnionOperatorWithPlus,
    NDashArySquareIntersectionOperator,
    NDashArySquareUnionOperator,
    TwoLogicalAndOperator,
    TwoLogicalOrOperator,
    NDashAryTimesOperator,
    ModuloTwoSum,
    SummationWithIntegral,
    QuadrupleIntegralOperator,
    FinitePartIntegral,
    IntegralWithDoubleStroke,
    IntegralAverageWithSlash,
    CirculationFunction,
    AnticlockwiseIntegration,
    LineIntegrationWithRectangularPathAroundPole,
    LineIntegrationWithSemicircularPathAroundPole,
    LineIntegrationNotIncludingThePole,
    IntegralAroundAPointOperator,
    QuaternionIntegralOperator,
    IntegralWithLeftwardsArrowWithHook,
    IntegralWithTimesSign,
    IntegralWithIntersection,
    IntegralWithUnion,
    IntegralWithOverbar,
    IntegralWithUnderbar,
    Join,
    LargeLeftTriangleOperator,
    ZNotationSchemaComposition,
    ZNotationSchemaPiping,
    ZNotationSchemaProjection,
    PlusSignWithSmallCircleAbove,
    PlusSignWithCircumflexAccentAbove,
    PlusSignWithTildeAbove,
    PlusSignWithDotBelow,
    PlusSignWithTildeBelow,
    PlusSignWithSubscriptTwo,
    PlusSignWithBlackTriangle,
    MinusSignWithCommaAbove,
    MinusSignWithDotBelow,
    MinusSignWithFallingDots,
    MinusSignWithRisingDots,
    PlusSignInLeftHalfCircle,
    PlusSignInRightHalfCircle,
    VectorOrCrossProduct,
    MultiplicationSignWithDotAbove,
    MultiplicationSignWithUnderbar,
    SemidirectProductWithBottomClosed,
    SmashProduct,
    MultiplicationSignInLeftHalfCircle,
    MultiplicationSignInRightHalfCircle,
    CircledMultiplicationSignWithCircumflexAccent,
    MultiplicationSignInDoubleCircle,
    CircledDivisionSign,
    PlusSignInTriangle,
    MinusSignInTriangle,
    MultiplicationSignInTriangle,
    InteriorProduct,
    RighthandInteriorProduct,
    ZNotationRelationalComposition,
    AmalgamationOrCoproduct,
    IntersectionWithDot,
    UnionWithMinusSign,
    UnionWithOverbar,
    IntersectionWithOverbar,
    IntersectionWithLogicalAnd,
    UnionWithLogicalOr,
    UnionAboveIntersection,
    IntersectionAboveUnion,
    UnionAboveBarAboveIntersection,
    IntersectionAboveBarAboveUnion,
    UnionBesideAndJoinedWithUnion,
    IntersectionBesideAndJoinedWithIntersection,
    ClosedUnionWithSerifs,
    ClosedIntersectionWithSerifs,
    DoubleSquareIntersection,
    DoubleSquareUnion,
    ClosedUnionWithSerifsAndSmashProduct,
    LogicalAndWithDotAbove,
    LogicalOrWithDotAbove,
    DoubleLogicalAnd,
    DoubleLogicalOr,
    TwoIntersectingLogicalAnd,
    TwoIntersectingLogicalOr,
    SlopingLargeOr,
    SlopingLargeAnd,
    LogicalOrOverlappingLogicalAnd,
    LogicalAndWithMiddleStem,
    LogicalOrWithMiddleStem,
    LogicalAndWithHorizontalDash,
    LogicalOrWithHorizontalDash,
    LogicalAndWithDoubleOverbar,
    LogicalAndWithUnderbar,
    LogicalAndWithDoubleUnderbar,
    SmallVeeWithUnderbar,
    LogicalOrWithDoubleOverbar,
    LogicalOrWithDoubleUnderbar,
    ZNotationDomainAntirestriction,
    ZNotationRangeAntirestriction,
    EqualsSignWithDotBelow,
    IdenticalWithDotAbove,
    TripleHorizontalBarWithDoubleVerticalStroke,
    TripleHorizontalBarWithTripleVerticalStroke,
    TildeOperatorWithDotAbove,
    TildeOperatorWithRisingDots,
    SimilarMinusSimilar,
    CongruentWithDotAbove,
    EqualsWithAsterisk,
    AlmostEqualToWithCircumflexAccent,
    ApproximatelyEqualOrEqualTo,
    EqualsSignAbovePlusSign,
    PlusSignAboveEqualsSign,
    EqualsSignAboveTildeOperator,
    DoubleColonEqual,
    TwoConsecutiveEqualsSigns,
    ThreeConsecutiveEqualsSigns,
    EqualsSignWithTwoDotsAboveAndTwoDotsBelow,
    EquivalentWithFourDotsAbove,
    LessDashThanWithCircleInside,
    GreaterDashThanWithCircleInside,
    LessDashThanWithQuestionMarkAbove,
    GreaterDashThanWithQuestionMarkAbove,
    LessDashThanOrSlantedEqualTo,
    GreaterDashThanOrSlantedEqualTo,
    LessDashThanOrSlantedEqualToWithDotInside,
    GreaterDashThanOrSlantedEqualToWithDotInside,
    LessDashThanOrSlantedEqualToWithDotAbove,
    GreaterDashThanOrSlantedEqualToWithDotAbove,
    LessDashThanOrSlantedEqualToWithDotAboveRight,
    GreaterDashThanOrSlantedEqualToWithDotAboveLeft,
    LessDashThanOrApproximate,
    GreaterDashThanOrApproximate,
    LessDashThanAndSingleDashLineNotEqualTo,
    GreaterDashThanAndSingleDashLineNotEqualTo,
    LessDashThanAndNotApproximate,
    GreaterDashThanAndNotApproximate,
    LessDashThanAboveDoubleDashLineEqualAboveGreaterDashThan,
    GreaterDashThanAboveDoubleDashLineEqualAboveLessDashThan,
    LessDashThanAboveSimilarOrEqual,
    GreaterDashThanAboveSimilarOrEqual,
    LessDashThanAboveSimilarAboveGreaterDashThan,
    GreaterDashThanAboveSimilarAboveLessDashThan,
    LessDashThanAboveGreaterDashThanAboveDoubleDashLineEqual,
    GreaterDashThanAboveLessDashThanAboveDoubleDashLineEqual,
    LessDashThanAboveSlantedEqualAboveGreaterDashThanAboveSlantedEqual,
    GreaterDashThanAboveSlantedEqualAboveLessDashThanAboveSlantedEqual,
    SlantedEqualToOrLessDashThan,
    SlantedEqualToOrGreaterDashThan,
    SlantedEqualToOrLessDashThanWithDotInside,
    SlantedEqualToOrGreaterDashThanWithDotInside,
    DoubleDashLineEqualToOrLessDashThan,
    DoubleDashLineEqualToOrGreaterDashThan,
    DoubleDashLineSlantedEqualToOrLessDashThan,
    DoubleDashLineSlantedEqualToOrGreaterDashThan,
    SimilarOrLessDashThan,
    SimilarOrGreaterDashThan,
    SimilarAboveLessDashThanAboveEqualsSign,
    SimilarAboveGreaterDashThanAboveEqualsSign,
    DoubleNestedLessDashThan,
    DoubleNestedGreaterDashThan,
    DoubleNestedLessDashThanWithUnderbar,
    GreaterDashThanOverlappingLessDashThan,
    GreaterDashThanBesideLessDashThan,
    LessDashThanClosedByCurve,
    GreaterDashThanClosedByCurve,
    LessDashThanClosedByCurveAboveSlantedEqual,
    GreaterDashThanClosedByCurveAboveSlantedEqual,
    SmallerThan,
    LargerThan,
    SmallerThanOrEqualTo,
    LargerThanOrEqualTo,
    EqualsSignWithBumpyAbove,
    PrecedesAboveSingleDashLineEqualsSign,
    SucceedsAboveSingleDashLineEqualsSign,
    PrecedesAboveSingleDashLineNotEqualTo,
    SucceedsAboveSingleDashLineNotEqualTo,
    PrecedesAboveEqualsSign,
    SucceedsAboveEqualsSign,
    PrecedesAboveNotEqualTo,
    SucceedsAboveNotEqualTo,
    PrecedesAboveAlmostEqualTo,
    SucceedsAboveAlmostEqualTo,
    PrecedesAboveNotAlmostEqualTo,
    SucceedsAboveNotAlmostEqualTo,
    DoublePrecedes,
    DoubleSucceeds,
    SubsetWithDot,
    SupersetWithDot,
    SubsetWithPlusSignBelow,
    SupersetWithPlusSignBelow,
    SubsetWithMultiplicationSignBelow,
    SupersetWithMultiplicationSignBelow,
    SubsetOfOrEqualToWithDotAbove,
    SupersetOfOrEqualToWithDotAbove,
    SubsetOfAboveEqualsSign,
    SupersetOfAboveEqualsSign,
    SubsetOfAboveTildeOperator,
    SupersetOfAboveTildeOperator,
    SubsetOfAboveAlmostEqualTo,
    SupersetOfAboveAlmostEqualTo,
    SubsetOfAboveNotEqualTo,
    SupersetOfAboveNotEqualTo,
    SquareLeftOpenBoxOperator,
    SquareRightOpenBoxOperator,
    ClosedSubset,
    ClosedSuperset,
    ClosedSubsetOrEqualTo,
    ClosedSupersetOrEqualTo,
    SubsetAboveSuperset,
    SupersetAboveSubset,
    SubsetAboveSubset,
    SupersetAboveSuperset,
    SupersetBesideSubset,
    SupersetBesideAndJoinedByDashWithSubset,
    ElementOfOpeningDownwards,
    PitchforkWithTeeTop,
    TransversalIntersection,
    Forking,
    Nonforking,
    ShortLeftTack,
    ShortDownTack,
    ShortUpTack,
    PerpendicularWithS,
    VerticalBarTripleRightTurnstile,
    DoubleVerticalBarLeftTurnstile,
    VerticalBarDoubleLeftTurnstile,
    DoubleVerticalBarDoubleLeftTurnstile,
    LongDashFromLeftMemberOfDoubleVertical,
    ShortDownTackWithOverbar,
    ShortUpTackWithUnderbar,
    ShortUpTackAboveShortDownTack,
    DoubleDownTack,
    DoubleUpTack,
    DoubleStrokeNotSign,
    ReversedDoubleStrokeNotSign,
    DoesNotDivideWithReversedNegationSlash,
    VerticalLineWithCircleAbove,
    VerticalLineWithCircleBelow,
    DownTackWithCircleBelow,
    ParallelWithHorizontalStroke,
    ParallelWithTildeOperator,
    TripleVerticalBarBinaryRelation,
    TripleVerticalBarWithHorizontalStroke,
    TripleColonOperator,
    TripleNestedLessDashThan,
    TripleNestedGreaterDashThan,
    DoubleDashLineSlantedLessDashThanOrEqualTo,
    DoubleDashLineSlantedGreaterDashThanOrEqualTo,
    TripleSolidusBinaryRelation,
    LargeTripleVerticalBarOperator,
    DoubleSolidusOperator,
    WhiteVerticalBar,
}

An enum to represent all characters in the SupplementalMathematicalOperators block.

Variants
NDashAryCircledDotOperator

\u{2a00}: '⨀'

NDashAryCircledPlusOperator

\u{2a01}: '⨁'

NDashAryCircledTimesOperator

\u{2a02}: '⨂'

NDashAryUnionOperatorWithDot

\u{2a03}: '⨃'

NDashAryUnionOperatorWithPlus

\u{2a04}: '⨄'

NDashArySquareIntersectionOperator

\u{2a05}: '⨅'

NDashArySquareUnionOperator

\u{2a06}: '⨆'

TwoLogicalAndOperator

\u{2a07}: '⨇'

TwoLogicalOrOperator

\u{2a08}: '⨈'

NDashAryTimesOperator

\u{2a09}: '⨉'

ModuloTwoSum

\u{2a0a}: '⨊'

SummationWithIntegral

\u{2a0b}: '⨋'

QuadrupleIntegralOperator

\u{2a0c}: '⨌'

FinitePartIntegral

\u{2a0d}: '⨍'

IntegralWithDoubleStroke

\u{2a0e}: '⨎'

IntegralAverageWithSlash

\u{2a0f}: '⨏'

CirculationFunction

\u{2a10}: '⨐'

AnticlockwiseIntegration

\u{2a11}: '⨑'

LineIntegrationWithRectangularPathAroundPole

\u{2a12}: '⨒'

LineIntegrationWithSemicircularPathAroundPole

\u{2a13}: '⨓'

LineIntegrationNotIncludingThePole

\u{2a14}: '⨔'

IntegralAroundAPointOperator

\u{2a15}: '⨕'

QuaternionIntegralOperator

\u{2a16}: '⨖'

IntegralWithLeftwardsArrowWithHook

\u{2a17}: '⨗'

IntegralWithTimesSign

\u{2a18}: '⨘'

IntegralWithIntersection

\u{2a19}: '⨙'

IntegralWithUnion

\u{2a1a}: '⨚'

IntegralWithOverbar

\u{2a1b}: '⨛'

IntegralWithUnderbar

\u{2a1c}: '⨜'

Join

\u{2a1d}: '⨝'

LargeLeftTriangleOperator

\u{2a1e}: '⨞'

ZNotationSchemaComposition

\u{2a1f}: '⨟'

ZNotationSchemaPiping

\u{2a20}: '⨠'

ZNotationSchemaProjection

\u{2a21}: '⨡'

PlusSignWithSmallCircleAbove

\u{2a22}: '⨢'

PlusSignWithCircumflexAccentAbove

\u{2a23}: '⨣'

PlusSignWithTildeAbove

\u{2a24}: '⨤'

PlusSignWithDotBelow

\u{2a25}: '⨥'

PlusSignWithTildeBelow

\u{2a26}: '⨦'

PlusSignWithSubscriptTwo

\u{2a27}: '⨧'

PlusSignWithBlackTriangle

\u{2a28}: '⨨'

MinusSignWithCommaAbove

\u{2a29}: '⨩'

MinusSignWithDotBelow

\u{2a2a}: '⨪'

MinusSignWithFallingDots

\u{2a2b}: '⨫'

MinusSignWithRisingDots

\u{2a2c}: '⨬'

PlusSignInLeftHalfCircle

\u{2a2d}: '⨭'

PlusSignInRightHalfCircle

\u{2a2e}: '⨮'

VectorOrCrossProduct

\u{2a2f}: '⨯'

MultiplicationSignWithDotAbove

\u{2a30}: '⨰'

MultiplicationSignWithUnderbar

\u{2a31}: '⨱'

SemidirectProductWithBottomClosed

\u{2a32}: '⨲'

SmashProduct

\u{2a33}: '⨳'

MultiplicationSignInLeftHalfCircle

\u{2a34}: '⨴'

MultiplicationSignInRightHalfCircle

\u{2a35}: '⨵'

CircledMultiplicationSignWithCircumflexAccent

\u{2a36}: '⨶'

MultiplicationSignInDoubleCircle

\u{2a37}: '⨷'

CircledDivisionSign

\u{2a38}: '⨸'

PlusSignInTriangle

\u{2a39}: '⨹'

MinusSignInTriangle

\u{2a3a}: '⨺'

MultiplicationSignInTriangle

\u{2a3b}: '⨻'

InteriorProduct

\u{2a3c}: '⨼'

RighthandInteriorProduct

\u{2a3d}: '⨽'

ZNotationRelationalComposition

\u{2a3e}: '⨾'

AmalgamationOrCoproduct

\u{2a3f}: '⨿'

IntersectionWithDot

\u{2a40}: '⩀'

UnionWithMinusSign

\u{2a41}: '⩁'

UnionWithOverbar

\u{2a42}: '⩂'

IntersectionWithOverbar

\u{2a43}: '⩃'

IntersectionWithLogicalAnd

\u{2a44}: '⩄'

UnionWithLogicalOr

\u{2a45}: '⩅'

UnionAboveIntersection

\u{2a46}: '⩆'

IntersectionAboveUnion

\u{2a47}: '⩇'

UnionAboveBarAboveIntersection

\u{2a48}: '⩈'

IntersectionAboveBarAboveUnion

\u{2a49}: '⩉'

UnionBesideAndJoinedWithUnion

\u{2a4a}: '⩊'

IntersectionBesideAndJoinedWithIntersection

\u{2a4b}: '⩋'

ClosedUnionWithSerifs

\u{2a4c}: '⩌'

ClosedIntersectionWithSerifs

\u{2a4d}: '⩍'

DoubleSquareIntersection

\u{2a4e}: '⩎'

DoubleSquareUnion

\u{2a4f}: '⩏'

ClosedUnionWithSerifsAndSmashProduct

\u{2a50}: '⩐'

LogicalAndWithDotAbove

\u{2a51}: '⩑'

LogicalOrWithDotAbove

\u{2a52}: '⩒'

DoubleLogicalAnd

\u{2a53}: '⩓'

DoubleLogicalOr

\u{2a54}: '⩔'

TwoIntersectingLogicalAnd

\u{2a55}: '⩕'

TwoIntersectingLogicalOr

\u{2a56}: '⩖'

SlopingLargeOr

\u{2a57}: '⩗'

SlopingLargeAnd

\u{2a58}: '⩘'

LogicalOrOverlappingLogicalAnd

\u{2a59}: '⩙'

LogicalAndWithMiddleStem

\u{2a5a}: '⩚'

LogicalOrWithMiddleStem

\u{2a5b}: '⩛'

LogicalAndWithHorizontalDash

\u{2a5c}: '⩜'

LogicalOrWithHorizontalDash

\u{2a5d}: '⩝'

LogicalAndWithDoubleOverbar

\u{2a5e}: '⩞'

LogicalAndWithUnderbar

\u{2a5f}: '⩟'

LogicalAndWithDoubleUnderbar

\u{2a60}: '⩠'

SmallVeeWithUnderbar

\u{2a61}: '⩡'

LogicalOrWithDoubleOverbar

\u{2a62}: '⩢'

LogicalOrWithDoubleUnderbar

\u{2a63}: '⩣'

ZNotationDomainAntirestriction

\u{2a64}: '⩤'

ZNotationRangeAntirestriction

\u{2a65}: '⩥'

EqualsSignWithDotBelow

\u{2a66}: '⩦'

IdenticalWithDotAbove

\u{2a67}: '⩧'

TripleHorizontalBarWithDoubleVerticalStroke

\u{2a68}: '⩨'

TripleHorizontalBarWithTripleVerticalStroke

\u{2a69}: '⩩'

TildeOperatorWithDotAbove

\u{2a6a}: '⩪'

TildeOperatorWithRisingDots

\u{2a6b}: '⩫'

SimilarMinusSimilar

\u{2a6c}: '⩬'

CongruentWithDotAbove

\u{2a6d}: '⩭'

EqualsWithAsterisk

\u{2a6e}: '⩮'

AlmostEqualToWithCircumflexAccent

\u{2a6f}: '⩯'

ApproximatelyEqualOrEqualTo

\u{2a70}: '⩰'

EqualsSignAbovePlusSign

\u{2a71}: '⩱'

PlusSignAboveEqualsSign

\u{2a72}: '⩲'

EqualsSignAboveTildeOperator

\u{2a73}: '⩳'

DoubleColonEqual

\u{2a74}: '⩴'

TwoConsecutiveEqualsSigns

\u{2a75}: '⩵'

ThreeConsecutiveEqualsSigns

\u{2a76}: '⩶'

EqualsSignWithTwoDotsAboveAndTwoDotsBelow

\u{2a77}: '⩷'

EquivalentWithFourDotsAbove

\u{2a78}: '⩸'

LessDashThanWithCircleInside

\u{2a79}: '⩹'

GreaterDashThanWithCircleInside

\u{2a7a}: '⩺'

LessDashThanWithQuestionMarkAbove

\u{2a7b}: '⩻'

GreaterDashThanWithQuestionMarkAbove

\u{2a7c}: '⩼'

LessDashThanOrSlantedEqualTo

\u{2a7d}: '⩽'

GreaterDashThanOrSlantedEqualTo

\u{2a7e}: '⩾'

LessDashThanOrSlantedEqualToWithDotInside

\u{2a7f}: '⩿'

GreaterDashThanOrSlantedEqualToWithDotInside

\u{2a80}: '⪀'

LessDashThanOrSlantedEqualToWithDotAbove

\u{2a81}: '⪁'

GreaterDashThanOrSlantedEqualToWithDotAbove

\u{2a82}: '⪂'

LessDashThanOrSlantedEqualToWithDotAboveRight

\u{2a83}: '⪃'

GreaterDashThanOrSlantedEqualToWithDotAboveLeft

\u{2a84}: '⪄'

LessDashThanOrApproximate

\u{2a85}: '⪅'

GreaterDashThanOrApproximate

\u{2a86}: '⪆'

LessDashThanAndSingleDashLineNotEqualTo

\u{2a87}: '⪇'

GreaterDashThanAndSingleDashLineNotEqualTo

\u{2a88}: '⪈'

LessDashThanAndNotApproximate

\u{2a89}: '⪉'

GreaterDashThanAndNotApproximate

\u{2a8a}: '⪊'

LessDashThanAboveDoubleDashLineEqualAboveGreaterDashThan

\u{2a8b}: '⪋'

GreaterDashThanAboveDoubleDashLineEqualAboveLessDashThan

\u{2a8c}: '⪌'

LessDashThanAboveSimilarOrEqual

\u{2a8d}: '⪍'

GreaterDashThanAboveSimilarOrEqual

\u{2a8e}: '⪎'

LessDashThanAboveSimilarAboveGreaterDashThan

\u{2a8f}: '⪏'

GreaterDashThanAboveSimilarAboveLessDashThan

\u{2a90}: '⪐'

LessDashThanAboveGreaterDashThanAboveDoubleDashLineEqual

\u{2a91}: '⪑'

GreaterDashThanAboveLessDashThanAboveDoubleDashLineEqual

\u{2a92}: '⪒'

LessDashThanAboveSlantedEqualAboveGreaterDashThanAboveSlantedEqual

\u{2a93}: '⪓'

GreaterDashThanAboveSlantedEqualAboveLessDashThanAboveSlantedEqual

\u{2a94}: '⪔'

SlantedEqualToOrLessDashThan

\u{2a95}: '⪕'

SlantedEqualToOrGreaterDashThan

\u{2a96}: '⪖'

SlantedEqualToOrLessDashThanWithDotInside

\u{2a97}: '⪗'

SlantedEqualToOrGreaterDashThanWithDotInside

\u{2a98}: '⪘'

DoubleDashLineEqualToOrLessDashThan

\u{2a99}: '⪙'

DoubleDashLineEqualToOrGreaterDashThan

\u{2a9a}: '⪚'

DoubleDashLineSlantedEqualToOrLessDashThan

\u{2a9b}: '⪛'

DoubleDashLineSlantedEqualToOrGreaterDashThan

\u{2a9c}: '⪜'

SimilarOrLessDashThan

\u{2a9d}: '⪝'

SimilarOrGreaterDashThan

\u{2a9e}: '⪞'

SimilarAboveLessDashThanAboveEqualsSign

\u{2a9f}: '⪟'

SimilarAboveGreaterDashThanAboveEqualsSign

\u{2aa0}: '⪠'

DoubleNestedLessDashThan

\u{2aa1}: '⪡'

DoubleNestedGreaterDashThan

\u{2aa2}: '⪢'

DoubleNestedLessDashThanWithUnderbar

\u{2aa3}: '⪣'

GreaterDashThanOverlappingLessDashThan

\u{2aa4}: '⪤'

GreaterDashThanBesideLessDashThan

\u{2aa5}: '⪥'

LessDashThanClosedByCurve

\u{2aa6}: '⪦'

GreaterDashThanClosedByCurve

\u{2aa7}: '⪧'

LessDashThanClosedByCurveAboveSlantedEqual

\u{2aa8}: '⪨'

GreaterDashThanClosedByCurveAboveSlantedEqual

\u{2aa9}: '⪩'

SmallerThan

\u{2aaa}: '⪪'

LargerThan

\u{2aab}: '⪫'

SmallerThanOrEqualTo

\u{2aac}: '⪬'

LargerThanOrEqualTo

\u{2aad}: '⪭'

EqualsSignWithBumpyAbove

\u{2aae}: '⪮'

PrecedesAboveSingleDashLineEqualsSign

\u{2aaf}: '⪯'

SucceedsAboveSingleDashLineEqualsSign

\u{2ab0}: '⪰'

PrecedesAboveSingleDashLineNotEqualTo

\u{2ab1}: '⪱'

SucceedsAboveSingleDashLineNotEqualTo

\u{2ab2}: '⪲'

PrecedesAboveEqualsSign

\u{2ab3}: '⪳'

SucceedsAboveEqualsSign

\u{2ab4}: '⪴'

PrecedesAboveNotEqualTo

\u{2ab5}: '⪵'

SucceedsAboveNotEqualTo

\u{2ab6}: '⪶'

PrecedesAboveAlmostEqualTo

\u{2ab7}: '⪷'

SucceedsAboveAlmostEqualTo

\u{2ab8}: '⪸'

PrecedesAboveNotAlmostEqualTo

\u{2ab9}: '⪹'

SucceedsAboveNotAlmostEqualTo

\u{2aba}: '⪺'

DoublePrecedes

\u{2abb}: '⪻'

DoubleSucceeds

\u{2abc}: '⪼'

SubsetWithDot

\u{2abd}: '⪽'

SupersetWithDot

\u{2abe}: '⪾'

SubsetWithPlusSignBelow

\u{2abf}: '⪿'

SupersetWithPlusSignBelow

\u{2ac0}: '⫀'

SubsetWithMultiplicationSignBelow

\u{2ac1}: '⫁'

SupersetWithMultiplicationSignBelow

\u{2ac2}: '⫂'

SubsetOfOrEqualToWithDotAbove

\u{2ac3}: '⫃'

SupersetOfOrEqualToWithDotAbove

\u{2ac4}: '⫄'

SubsetOfAboveEqualsSign

\u{2ac5}: '⫅'

SupersetOfAboveEqualsSign

\u{2ac6}: '⫆'

SubsetOfAboveTildeOperator

\u{2ac7}: '⫇'

SupersetOfAboveTildeOperator

\u{2ac8}: '⫈'

SubsetOfAboveAlmostEqualTo

\u{2ac9}: '⫉'

SupersetOfAboveAlmostEqualTo

\u{2aca}: '⫊'

SubsetOfAboveNotEqualTo

\u{2acb}: '⫋'

SupersetOfAboveNotEqualTo

\u{2acc}: '⫌'

SquareLeftOpenBoxOperator

\u{2acd}: '⫍'

SquareRightOpenBoxOperator

\u{2ace}: '⫎'

ClosedSubset

\u{2acf}: '⫏'

ClosedSuperset

\u{2ad0}: '⫐'

ClosedSubsetOrEqualTo

\u{2ad1}: '⫑'

ClosedSupersetOrEqualTo

\u{2ad2}: '⫒'

SubsetAboveSuperset

\u{2ad3}: '⫓'

SupersetAboveSubset

\u{2ad4}: '⫔'

SubsetAboveSubset

\u{2ad5}: '⫕'

SupersetAboveSuperset

\u{2ad6}: '⫖'

SupersetBesideSubset

\u{2ad7}: '⫗'

SupersetBesideAndJoinedByDashWithSubset

\u{2ad8}: '⫘'

ElementOfOpeningDownwards

\u{2ad9}: '⫙'

PitchforkWithTeeTop

\u{2ada}: '⫚'

TransversalIntersection

\u{2adb}: '⫛'

Forking

\u{2adc}: '⫝̸'

Nonforking

\u{2add}: '⫝'

ShortLeftTack

\u{2ade}: '⫞'

ShortDownTack

\u{2adf}: '⫟'

ShortUpTack

\u{2ae0}: '⫠'

PerpendicularWithS

\u{2ae1}: '⫡'

VerticalBarTripleRightTurnstile

\u{2ae2}: '⫢'

DoubleVerticalBarLeftTurnstile

\u{2ae3}: '⫣'

VerticalBarDoubleLeftTurnstile

\u{2ae4}: '⫤'

DoubleVerticalBarDoubleLeftTurnstile

\u{2ae5}: '⫥'

LongDashFromLeftMemberOfDoubleVertical

\u{2ae6}: '⫦'

ShortDownTackWithOverbar

\u{2ae7}: '⫧'

ShortUpTackWithUnderbar

\u{2ae8}: '⫨'

ShortUpTackAboveShortDownTack

\u{2ae9}: '⫩'

DoubleDownTack

\u{2aea}: '⫪'

DoubleUpTack

\u{2aeb}: '⫫'

DoubleStrokeNotSign

\u{2aec}: '⫬'

ReversedDoubleStrokeNotSign

\u{2aed}: '⫭'

DoesNotDivideWithReversedNegationSlash

\u{2aee}: '⫮'

VerticalLineWithCircleAbove

\u{2aef}: '⫯'

VerticalLineWithCircleBelow

\u{2af0}: '⫰'

DownTackWithCircleBelow

\u{2af1}: '⫱'

ParallelWithHorizontalStroke

\u{2af2}: '⫲'

ParallelWithTildeOperator

\u{2af3}: '⫳'

TripleVerticalBarBinaryRelation

\u{2af4}: '⫴'

TripleVerticalBarWithHorizontalStroke

\u{2af5}: '⫵'

TripleColonOperator

\u{2af6}: '⫶'

TripleNestedLessDashThan

\u{2af7}: '⫷'

TripleNestedGreaterDashThan

\u{2af8}: '⫸'

DoubleDashLineSlantedLessDashThanOrEqualTo

\u{2af9}: '⫹'

DoubleDashLineSlantedGreaterDashThanOrEqualTo

\u{2afa}: '⫺'

TripleSolidusBinaryRelation

\u{2afb}: '⫻'

LargeTripleVerticalBarOperator

\u{2afc}: '⫼'

DoubleSolidusOperator

\u{2afd}: '⫽'

WhiteVerticalBar

\u{2afe}: '⫾'

Methods

impl SupplementalMathematicalOperators

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case