Enum unicode_types::supplemental_punctuation::SupplementalPunctuation

pub enum SupplementalPunctuation {
    RightAngleSubstitutionMarker,
    RightAngleDottedSubstitutionMarker,
    LeftSubstitutionBracket,
    RightSubstitutionBracket,
    LeftDottedSubstitutionBracket,
    RightDottedSubstitutionBracket,
    RaisedInterpolationMarker,
    RaisedDottedInterpolationMarker,
    DottedTranspositionMarker,
    LeftTranspositionBracket,
    RightTranspositionBracket,
    RaisedSquare,
    LeftRaisedOmissionBracket,
    RightRaisedOmissionBracket,
    EditorialCoronis,
    Paragraphos,
    ForkedParagraphos,
    ReversedForkedParagraphos,
    Hypodiastole,
    DottedObelos,
    DownwardsAncora,
    UpwardsAncora,
    DottedRightDashPointingAngle,
    DoubleObliqueHyphen,
    InvertedInterrobang,
    PalmBranch,
    HyphenWithDiaeresis,
    TildeWithRingAbove,
    LeftLowParaphraseBracket,
    RightLowParaphraseBracket,
    TildeWithDotAbove,
    TildeWithDotBelow,
    LeftVerticalBarWithQuill,
    RightVerticalBarWithQuill,
    TopLeftHalfBracket,
    TopRightHalfBracket,
    BottomLeftHalfBracket,
    BottomRightHalfBracket,
    LeftSidewaysUBracket,
    RightSidewaysUBracket,
    LeftDoubleParenthesis,
    RightDoubleParenthesis,
    TwoDotsOverOneDotPunctuation,
    OneDotOverTwoDotsPunctuation,
    SquaredFourDotPunctuation,
    FiveDotMark,
    ReversedQuestionMark,
    VerticalTilde,
    RingPoint,
    WordSeparatorMiddleDot,
    TurnedComma,
    RaisedDot,
    RaisedComma,
    TurnedSemicolon,
    DaggerWithLeftGuard,
    DaggerWithRightGuard,
    TurnedDagger,
    TopHalfSectionSign,
    TwoDashEmDash,
    ThreeDashEmDash,
    StenographicFullStop,
    VerticalSixDots,
    WigglyVerticalLine,
    Capitulum,
    DoubleHyphen,
    ReversedComma,
    DoubleLowDashReversedDash9QuotationMark,
    DashWithLeftUpturn,
    DoubleSuspensionMark,
    InvertedLowKavyka,
    InvertedLowKavykaWithKavykaAbove,
    LowKavyka,
    LowKavykaWithDot,
    DoubleStackedComma,
    DottedSolidus,
    TripleDagger,
    MedievalComma,
    ParagraphusMark,
    PunctusElevatusMark,
    CornishVerseDivider,
}

An enum to represent all characters in the SupplementalPunctuation block.

Variants
RightAngleSubstitutionMarker

\u{2e00}: '⸀'

RightAngleDottedSubstitutionMarker

\u{2e01}: '⸁'

LeftSubstitutionBracket

\u{2e02}: '⸂'

RightSubstitutionBracket

\u{2e03}: '⸃'

LeftDottedSubstitutionBracket

\u{2e04}: '⸄'

RightDottedSubstitutionBracket

\u{2e05}: '⸅'

RaisedInterpolationMarker

\u{2e06}: '⸆'

RaisedDottedInterpolationMarker

\u{2e07}: '⸇'

DottedTranspositionMarker

\u{2e08}: '⸈'

LeftTranspositionBracket

\u{2e09}: '⸉'

RightTranspositionBracket

\u{2e0a}: '⸊'

RaisedSquare

\u{2e0b}: '⸋'

LeftRaisedOmissionBracket

\u{2e0c}: '⸌'

RightRaisedOmissionBracket

\u{2e0d}: '⸍'

EditorialCoronis

\u{2e0e}: '⸎'

Paragraphos

\u{2e0f}: '⸏'

ForkedParagraphos

\u{2e10}: '⸐'

ReversedForkedParagraphos

\u{2e11}: '⸑'

Hypodiastole

\u{2e12}: '⸒'

DottedObelos

\u{2e13}: '⸓'

DownwardsAncora

\u{2e14}: '⸔'

UpwardsAncora

\u{2e15}: '⸕'

DottedRightDashPointingAngle

\u{2e16}: '⸖'

DoubleObliqueHyphen

\u{2e17}: '⸗'

InvertedInterrobang

\u{2e18}: '⸘'

PalmBranch

\u{2e19}: '⸙'

HyphenWithDiaeresis

\u{2e1a}: '⸚'

TildeWithRingAbove

\u{2e1b}: '⸛'

LeftLowParaphraseBracket

\u{2e1c}: '⸜'

RightLowParaphraseBracket

\u{2e1d}: '⸝'

TildeWithDotAbove

\u{2e1e}: '⸞'

TildeWithDotBelow

\u{2e1f}: '⸟'

LeftVerticalBarWithQuill

\u{2e20}: '⸠'

RightVerticalBarWithQuill

\u{2e21}: '⸡'

TopLeftHalfBracket

\u{2e22}: '⸢'

TopRightHalfBracket

\u{2e23}: '⸣'

BottomLeftHalfBracket

\u{2e24}: '⸤'

BottomRightHalfBracket

\u{2e25}: '⸥'

LeftSidewaysUBracket

\u{2e26}: '⸦'

RightSidewaysUBracket

\u{2e27}: '⸧'

LeftDoubleParenthesis

\u{2e28}: '⸨'

RightDoubleParenthesis

\u{2e29}: '⸩'

TwoDotsOverOneDotPunctuation

\u{2e2a}: '⸪'

OneDotOverTwoDotsPunctuation

\u{2e2b}: '⸫'

SquaredFourDotPunctuation

\u{2e2c}: '⸬'

FiveDotMark

\u{2e2d}: '⸭'

ReversedQuestionMark

\u{2e2e}: '⸮'

VerticalTilde

\u{2e2f}: 'ⸯ'

RingPoint

\u{2e30}: '⸰'

WordSeparatorMiddleDot

\u{2e31}: '⸱'

TurnedComma

\u{2e32}: '⸲'

RaisedDot

\u{2e33}: '⸳'

RaisedComma

\u{2e34}: '⸴'

TurnedSemicolon

\u{2e35}: '⸵'

DaggerWithLeftGuard

\u{2e36}: '⸶'

DaggerWithRightGuard

\u{2e37}: '⸷'

TurnedDagger

\u{2e38}: '⸸'

TopHalfSectionSign

\u{2e39}: '⸹'

TwoDashEmDash

\u{2e3a}: '⸺'

ThreeDashEmDash

\u{2e3b}: '⸻'

StenographicFullStop

\u{2e3c}: '⸼'

VerticalSixDots

\u{2e3d}: '⸽'

WigglyVerticalLine

\u{2e3e}: '⸾'

Capitulum

\u{2e3f}: '⸿'

DoubleHyphen

\u{2e40}: '⹀'

ReversedComma

\u{2e41}: '⹁'

DoubleLowDashReversedDash9QuotationMark

\u{2e42}: '⹂'

DashWithLeftUpturn

\u{2e43}: '⹃'

DoubleSuspensionMark

\u{2e44}: '⹄'

InvertedLowKavyka

\u{2e45}: '⹅'

InvertedLowKavykaWithKavykaAbove

\u{2e46}: '⹆'

LowKavyka

\u{2e47}: '⹇'

LowKavykaWithDot

\u{2e48}: '⹈'

DoubleStackedComma

\u{2e49}: '⹉'

DottedSolidus

\u{2e4a}: '⹊'

TripleDagger

\u{2e4b}: '⹋'

MedievalComma

\u{2e4c}: '⹌'

ParagraphusMark

\u{2e4d}: '⹍'

PunctusElevatusMark

\u{2e4e}: '⹎'

CornishVerseDivider

\u{2e4f}: '⹏'

Methods

impl SupplementalPunctuation

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case