Enum unicode_types::supplemental_symbols_and_pictographs::SupplementalSymbolsandPictographs

pub enum SupplementalSymbolsandPictographs {
    CircledCrossFormeeWithFourDots,
    CircledCrossFormeeWithTwoDots,
    CircledCrossFormee,
    LeftHalfCircleWithFourDots,
    LeftHalfCircleWithThreeDots,
    LeftHalfCircleWithTwoDots,
    LeftHalfCircleWithDot,
    LeftHalfCircle,
    DownwardFacingHook,
    DownwardFacingNotchedHook,
    DownwardFacingHookWithDot,
    DownwardFacingNotchedHookWithDot,
    WhiteHeart,
    BrownHeart,
    PinchingHand,
    ZipperDashMouthFace,
    MoneyDashMouthFace,
    FaceWithThermometer,
    NerdFace,
    ThinkingFace,
    FaceWithHeadDashBandage,
    RobotFace,
    HuggingFace,
    SignOfTheHorns,
    CallMeHand,
    RaisedBackOfHand,
    LeftDashFacingFist,
    RightDashFacingFist,
    Handshake,
    HandWithIndexAndMiddleFingersCrossed,
    ILoveYouHandSign,
    FaceWithCowboyHat,
    ClownFace,
    NauseatedFace,
    RollingOnTheFloorLaughing,
    DroolingFace,
    LyingFace,
    FacePalm,
    SneezingFace,
    FaceWithOneEyebrowRaised,
    GrinningFaceWithStarEyes,
    GrinningFaceWithOneLargeAndOneSmallEye,
    FaceWithFingerCoveringClosedLips,
    SeriousFaceWithSymbolsCoveringMouth,
    SmilingFaceWithSmilingEyesAndHandCoveringMouth,
    FaceWithOpenMouthVomiting,
    ShockedFaceWithExplodingHead,
    PregnantWoman,
    BreastDashFeeding,
    PalmsUpTogether,
    Selfie,
    Prince,
    ManInTuxedo,
    MotherChristmas,
    Shrug,
    PersonDoingCartwheel,
    Juggling,
    Fencer,
    ModernPentathlon,
    Wrestlers,
    WaterPolo,
    Handball,
    DivingMask,
    WiltedFlower,
    DrumWithDrumsticks,
    ClinkingGlasses,
    TumblerGlass,
    Spoon,
    GoalNet,
    Rifle,
    FirstPlaceMedal,
    SecondPlaceMedal,
    ThirdPlaceMedal,
    BoxingGlove,
    MartialArtsUniform,
    CurlingStone,
    LacrosseStickAndBall,
    Softball,
    FlyingDisc,
    Croissant,
    Avocado,
    Cucumber,
    Bacon,
    Potato,
    Carrot,
    BaguetteBread,
    GreenSalad,
    ShallowPanOfFood,
    StuffedFlatbread,
    Egg,
    GlassOfMilk,
    Peanuts,
    Kiwifruit,
    Pancakes,
    Dumpling,
    FortuneCookie,
    TakeoutBox,
    Chopsticks,
    BowlWithSpoon,
    CupWithStraw,
    Coconut,
    Broccoli,
    Pie,
    Pretzel,
    CutOfMeat,
    Sandwich,
    CannedFood,
    LeafyGreen,
    Mango,
    MoonCake,
    Bagel,
    SmilingFaceWithSmilingEyesAndThreeHearts,
    YawningFace,
    FaceWithPartyHornAndPartyHat,
    FaceWithUnevenEyesAndWavyMouth,
    OverheatedFace,
    FreezingFace,
    FaceWithPleadingEyes,
    Sari,
    LabCoat,
    Goggles,
    HikingBoot,
    FlatShoe,
    Crab,
    LionFace,
    Scorpion,
    Turkey,
    UnicornFace,
    Eagle,
    Duck,
    Bat,
    Shark,
    Owl,
    FoxFace,
    Butterfly,
    Deer,
    Gorilla,
    Lizard,
    Rhinoceros,
    Shrimp,
    Squid,
    GiraffeFace,
    ZebraFace,
    Hedgehog,
    Sauropod,
    TDashRex,
    Cricket,
    Kangaroo,
    Llama,
    Peacock,
    Hippopotamus,
    Parrot,
    Raccoon,
    Lobster,
    Mosquito,
    Microbe,
    Badger,
    Swan,
    Sloth,
    Otter,
    Orangutan,
    Skunk,
    Flamingo,
    Oyster,
    GuideDog,
    ProbingCane,
    EmojiComponentRedHair,
    EmojiComponentCurlyHair,
    EmojiComponentBald,
    EmojiComponentWhiteHair,
    Bone,
    Leg,
    Foot,
    Tooth,
    Superhero,
    Supervillain,
    SafetyVest,
    EarWithHearingAid,
    MotorizedWheelchair,
    ManualWheelchair,
    MechanicalArm,
    MechanicalLeg,
    CheeseWedge,
    Cupcake,
    SaltShaker,
    BeverageBox,
    Garlic,
    Onion,
    Falafel,
    Waffle,
    Butter,
    MateDrink,
    IceCube,
    StandingPerson,
    KneelingPerson,
    DeafPerson,
    FaceWithMonocle,
    Adult,
    Child,
    OlderAdult,
    BeardedPerson,
    PersonWithHeadscarf,
    PersonInSteamyRoom,
    PersonClimbing,
    PersonInLotusPosition,
    Mage,
    Fairy,
    Vampire,
    Merperson,
    Elf,
    Genie,
    Zombie,
    Brain,
    OrangeHeart,
    BilledCap,
    Scarf,
    Gloves,
    Coat,
    Socks,
    RedGiftEnvelope,
    Firecracker,
    JigsawPuzzlePiece,
    TestTube,
    PetriDish,
    DnaDoubleHelix,
    Compass,
    Abacus,
    FireExtinguisher,
    Toolbox,
    Brick,
    Magnet,
    Luggage,
    LotionBottle,
    SpoolOfThread,
    BallOfYarn,
    SafetyPin,
    TeddyBear,
    Broom,
    Basket,
    RollOfPaper,
    BarOfSoap,
    Sponge,
    Receipt,
}

An enum to represent all characters in the SupplementalSymbolsandPictographs block.

Variants
CircledCrossFormeeWithFourDots

\u{1f900}: 'ğŸ¤€'

CircledCrossFormeeWithTwoDots

\u{1f901}: 'ğŸ¤'

CircledCrossFormee

\u{1f902}: 'ğŸ¤‚'

LeftHalfCircleWithFourDots

\u{1f903}: 'ğŸ¤ƒ'

LeftHalfCircleWithThreeDots

\u{1f904}: 'ğŸ¤„'

LeftHalfCircleWithTwoDots

\u{1f905}: 'ğŸ¤…'

LeftHalfCircleWithDot

\u{1f906}: 'ğŸ¤†'

LeftHalfCircle

\u{1f907}: 'ğŸ¤‡'

DownwardFacingHook

\u{1f908}: 'ğŸ¤ˆ'

DownwardFacingNotchedHook

\u{1f909}: 'ğŸ¤‰'

DownwardFacingHookWithDot

\u{1f90a}: 'ğŸ¤Š'

DownwardFacingNotchedHookWithDot

\u{1f90b}: 'ğŸ¤‹'

WhiteHeart

\u{1f90d}: 'ğŸ¤'

BrownHeart

\u{1f90e}: 'ğŸ¤'

PinchingHand

\u{1f90f}: 'ğŸ¤'

ZipperDashMouthFace

\u{1f910}: 'ğŸ¤'

MoneyDashMouthFace

\u{1f911}: 'ğŸ¤‘'

FaceWithThermometer

\u{1f912}: 'ğŸ¤’'

NerdFace

\u{1f913}: 'ğŸ¤“'

ThinkingFace

\u{1f914}: 'ğŸ¤”'

FaceWithHeadDashBandage

\u{1f915}: 'ğŸ¤•'

RobotFace

\u{1f916}: 'ğŸ¤–'

HuggingFace

\u{1f917}: 'ğŸ¤—'

SignOfTheHorns

\u{1f918}: 'ğŸ¤˜'

CallMeHand

\u{1f919}: 'ğŸ¤™'

RaisedBackOfHand

\u{1f91a}: 'ğŸ¤š'

LeftDashFacingFist

\u{1f91b}: 'ğŸ¤›'

RightDashFacingFist

\u{1f91c}: 'ğŸ¤œ'

Handshake

\u{1f91d}: 'ğŸ¤'

HandWithIndexAndMiddleFingersCrossed

\u{1f91e}: 'ğŸ¤'

ILoveYouHandSign

\u{1f91f}: 'ğŸ¤Ÿ'

FaceWithCowboyHat

\u{1f920}: 'ğŸ¤ '

ClownFace

\u{1f921}: 'ğŸ¤¡'

NauseatedFace

\u{1f922}: 'ğŸ¤¢'

RollingOnTheFloorLaughing

\u{1f923}: 'ğŸ¤£'

DroolingFace

\u{1f924}: 'ğŸ¤¤'

LyingFace

\u{1f925}: 'ğŸ¤¥'

FacePalm

\u{1f926}: 'ğŸ¤¦'

SneezingFace

\u{1f927}: 'ğŸ¤'

FaceWithOneEyebrowRaised

\u{1f928}: 'ğŸ¤¨'

GrinningFaceWithStarEyes

\u{1f929}: 'ğŸ¤©'

GrinningFaceWithOneLargeAndOneSmallEye

\u{1f92a}: 'ğŸ¤ª'

FaceWithFingerCoveringClosedLips

\u{1f92b}: 'ğŸ¤«'

SeriousFaceWithSymbolsCoveringMouth

\u{1f92c}: 'ğŸ¤¬'

SmilingFaceWithSmilingEyesAndHandCoveringMouth

\u{1f92d}: 'ğŸ¤­'

FaceWithOpenMouthVomiting

\u{1f92e}: 'ğŸ¤®'

ShockedFaceWithExplodingHead

\u{1f92f}: 'ğŸ¤¯'

PregnantWoman

\u{1f930}: 'ğŸ¤°'

BreastDashFeeding

\u{1f931}: 'ğŸ¤±'

PalmsUpTogether

\u{1f932}: 'ğŸ¤²'

Selfie

\u{1f933}: 'ğŸ¤³'

Prince

\u{1f934}: 'ğŸ¤´'

ManInTuxedo

\u{1f935}: 'ğŸ¤µ'

MotherChristmas

\u{1f936}: 'ğŸ¤¶'

Shrug

\u{1f937}: 'ğŸ¤·'

PersonDoingCartwheel

\u{1f938}: 'ğŸ¤¸'

Juggling

\u{1f939}: 'ğŸ¤¹'

Fencer

\u{1f93a}: 'ğŸ¤º'

ModernPentathlon

\u{1f93b}: 'ğŸ¤»'

Wrestlers

\u{1f93c}: 'ğŸ¤¼'

WaterPolo

\u{1f93d}: 'ğŸ¤½'

Handball

\u{1f93e}: 'ğŸ¤¾'

DivingMask

\u{1f93f}: 'ğŸ¤¿'

WiltedFlower

\u{1f940}: 'ğŸ¥€'

DrumWithDrumsticks

\u{1f941}: 'ğŸ¥'

ClinkingGlasses

\u{1f942}: 'ğŸ¥‚'

TumblerGlass

\u{1f943}: 'ğŸ¥ƒ'

Spoon

\u{1f944}: 'ğŸ¥„'

GoalNet

\u{1f945}: 'ğŸ¥…'

Rifle

\u{1f946}: 'ğŸ¥†'

FirstPlaceMedal

\u{1f947}: 'ğŸ¥‡'

SecondPlaceMedal

\u{1f948}: 'ğŸ¥ˆ'

ThirdPlaceMedal

\u{1f949}: 'ğŸ¥‰'

BoxingGlove

\u{1f94a}: 'ğŸ¥Š'

MartialArtsUniform

\u{1f94b}: 'ğŸ¥‹'

CurlingStone

\u{1f94c}: 'ğŸ¥Œ'

LacrosseStickAndBall

\u{1f94d}: 'ğŸ¥'

Softball

\u{1f94e}: 'ğŸ¥'

FlyingDisc

\u{1f94f}: 'ğŸ¥'

Croissant

\u{1f950}: 'ğŸ¥'

Avocado

\u{1f951}: 'ğŸ¥‘'

Cucumber

\u{1f952}: 'ğŸ¥’'

Bacon

\u{1f953}: 'ğŸ¥“'

Potato

\u{1f954}: 'ğŸ¥”'

Carrot

\u{1f955}: 'ğŸ¥•'

BaguetteBread

\u{1f956}: 'ğŸ¥–'

GreenSalad

\u{1f957}: 'ğŸ¥—'

ShallowPanOfFood

\u{1f958}: 'ğŸ¥˜'

StuffedFlatbread

\u{1f959}: 'ğŸ¥™'

Egg

\u{1f95a}: 'ğŸ¥š'

GlassOfMilk

\u{1f95b}: 'ğŸ¥›'

Peanuts

\u{1f95c}: 'ğŸ¥œ'

Kiwifruit

\u{1f95d}: 'ğŸ¥'

Pancakes

\u{1f95e}: 'ğŸ¥'

Dumpling

\u{1f95f}: 'ğŸ¥Ÿ'

FortuneCookie

\u{1f960}: 'ğŸ¥ '

TakeoutBox

\u{1f961}: 'ğŸ¥¡'

Chopsticks

\u{1f962}: 'ğŸ¥¢'

BowlWithSpoon

\u{1f963}: 'ğŸ¥£'

CupWithStraw

\u{1f964}: 'ğŸ¥¤'

Coconut

\u{1f965}: 'ğŸ¥¥'

Broccoli

\u{1f966}: 'ğŸ¥¦'

Pie

\u{1f967}: 'ğŸ¥'

Pretzel

\u{1f968}: 'ğŸ¥¨'

CutOfMeat

\u{1f969}: 'ğŸ¥©'

Sandwich

\u{1f96a}: 'ğŸ¥ª'

CannedFood

\u{1f96b}: 'ğŸ¥«'

LeafyGreen

\u{1f96c}: 'ğŸ¥¬'

Mango

\u{1f96d}: 'ğŸ¥­'

MoonCake

\u{1f96e}: 'ğŸ¥®'

Bagel

\u{1f96f}: 'ğŸ¥¯'

SmilingFaceWithSmilingEyesAndThreeHearts

\u{1f970}: 'ğŸ¥°'

YawningFace

\u{1f971}: 'ğŸ¥±'

FaceWithPartyHornAndPartyHat

\u{1f973}: 'ğŸ¥³'

FaceWithUnevenEyesAndWavyMouth

\u{1f974}: 'ğŸ¥´'

OverheatedFace

\u{1f975}: 'ğŸ¥µ'

FreezingFace

\u{1f976}: 'ğŸ¥¶'

FaceWithPleadingEyes

\u{1f97a}: 'ğŸ¥º'

Sari

\u{1f97b}: 'ğŸ¥»'

LabCoat

\u{1f97c}: 'ğŸ¥¼'

Goggles

\u{1f97d}: 'ğŸ¥½'

HikingBoot

\u{1f97e}: 'ğŸ¥¾'

FlatShoe

\u{1f97f}: 'ğŸ¥¿'

Crab

\u{1f980}: 'ğŸ¦€'

LionFace

\u{1f981}: 'ğŸ¦'

Scorpion

\u{1f982}: 'ğŸ¦‚'

Turkey

\u{1f983}: 'ğŸ¦ƒ'

UnicornFace

\u{1f984}: 'ğŸ¦„'

Eagle

\u{1f985}: 'ğŸ¦…'

Duck

\u{1f986}: 'ğŸ¦†'

Bat

\u{1f987}: 'ğŸ¦‡'

Shark

\u{1f988}: 'ğŸ¦ˆ'

Owl

\u{1f989}: 'ğŸ¦‰'

FoxFace

\u{1f98a}: 'ğŸ¦Š'

Butterfly

\u{1f98b}: 'ğŸ¦‹'

Deer

\u{1f98c}: 'ğŸ¦Œ'

Gorilla

\u{1f98d}: 'ğŸ¦'

Lizard

\u{1f98e}: 'ğŸ¦'

Rhinoceros

\u{1f98f}: 'ğŸ¦'

Shrimp

\u{1f990}: 'ğŸ¦'

Squid

\u{1f991}: 'ğŸ¦‘'

GiraffeFace

\u{1f992}: 'ğŸ¦’'

ZebraFace

\u{1f993}: 'ğŸ¦“'

Hedgehog

\u{1f994}: 'ğŸ¦”'

Sauropod

\u{1f995}: 'ğŸ¦•'

TDashRex

\u{1f996}: 'ğŸ¦–'

Cricket

\u{1f997}: 'ğŸ¦—'

Kangaroo

\u{1f998}: 'ğŸ¦˜'

Llama

\u{1f999}: 'ğŸ¦™'

Peacock

\u{1f99a}: 'ğŸ¦š'

Hippopotamus

\u{1f99b}: 'ğŸ¦›'

Parrot

\u{1f99c}: 'ğŸ¦œ'

Raccoon

\u{1f99d}: 'ğŸ¦'

Lobster

\u{1f99e}: 'ğŸ¦'

Mosquito

\u{1f99f}: 'ğŸ¦Ÿ'

Microbe

\u{1f9a0}: 'ğŸ¦ '

Badger

\u{1f9a1}: 'ğŸ¦¡'

Swan

\u{1f9a2}: 'ğŸ¦¢'

Sloth

\u{1f9a5}: 'ğŸ¦¥'

Otter

\u{1f9a6}: 'ğŸ¦¦'

Orangutan

\u{1f9a7}: 'ğŸ¦'

Skunk

\u{1f9a8}: 'ğŸ¦¨'

Flamingo

\u{1f9a9}: 'ğŸ¦©'

Oyster

\u{1f9aa}: 'ğŸ¦ª'

GuideDog

\u{1f9ae}: 'ğŸ¦®'

ProbingCane

\u{1f9af}: 'ğŸ¦¯'

EmojiComponentRedHair

\u{1f9b0}: 'ğŸ¦°'

EmojiComponentCurlyHair

\u{1f9b1}: 'ğŸ¦±'

EmojiComponentBald

\u{1f9b2}: 'ğŸ¦²'

EmojiComponentWhiteHair

\u{1f9b3}: 'ğŸ¦³'

Bone

\u{1f9b4}: 'ğŸ¦´'

Leg

\u{1f9b5}: 'ğŸ¦µ'

Foot

\u{1f9b6}: 'ğŸ¦¶'

Tooth

\u{1f9b7}: 'ğŸ¦·'

Superhero

\u{1f9b8}: 'ğŸ¦¸'

Supervillain

\u{1f9b9}: 'ğŸ¦¹'

SafetyVest

\u{1f9ba}: 'ğŸ¦º'

EarWithHearingAid

\u{1f9bb}: 'ğŸ¦»'

MotorizedWheelchair

\u{1f9bc}: 'ğŸ¦¼'

ManualWheelchair

\u{1f9bd}: 'ğŸ¦½'

MechanicalArm

\u{1f9be}: 'ğŸ¦¾'

MechanicalLeg

\u{1f9bf}: 'ğŸ¦¿'

CheeseWedge

\u{1f9c0}: 'ğŸ€'

Cupcake

\u{1f9c1}: 'ğŸ'

SaltShaker

\u{1f9c2}: 'ğŸ‚'

BeverageBox

\u{1f9c3}: 'ğŸƒ'

Garlic

\u{1f9c4}: 'ğŸ„'

Onion

\u{1f9c5}: 'ğŸ…'

Falafel

\u{1f9c6}: 'ğŸ†'

Waffle

\u{1f9c7}: 'ğŸ‡'

Butter

\u{1f9c8}: 'ğŸˆ'

MateDrink

\u{1f9c9}: 'ğŸ‰'

IceCube

\u{1f9ca}: 'ğŸŠ'

StandingPerson

\u{1f9cd}: 'ğŸ'

KneelingPerson

\u{1f9ce}: 'ğŸ'

DeafPerson

\u{1f9cf}: 'ğŸ'

FaceWithMonocle

\u{1f9d0}: 'ğŸ'

Adult

\u{1f9d1}: 'ğŸ‘'

Child

\u{1f9d2}: 'ğŸ’'

OlderAdult

\u{1f9d3}: 'ğŸ“'

BeardedPerson

\u{1f9d4}: 'ğŸ”'

PersonWithHeadscarf

\u{1f9d5}: 'ğŸ•'

PersonInSteamyRoom

\u{1f9d6}: 'ğŸ–'

PersonClimbing

\u{1f9d7}: 'ğŸ—'

PersonInLotusPosition

\u{1f9d8}: 'ğŸ˜'

Mage

\u{1f9d9}: 'ğŸ™'

Fairy

\u{1f9da}: 'ğŸš'

Vampire

\u{1f9db}: 'ğŸ›'

Merperson

\u{1f9dc}: 'ğŸœ'

Elf

\u{1f9dd}: 'ğŸ'

Genie

\u{1f9de}: 'ğŸ'

Zombie

\u{1f9df}: 'ğŸŸ'

Brain

\u{1f9e0}: 'ğŸ '

OrangeHeart

\u{1f9e1}: 'ğŸ¡'

BilledCap

\u{1f9e2}: 'ğŸ¢'

Scarf

\u{1f9e3}: 'ğŸ£'

Gloves

\u{1f9e4}: 'ğŸ¤'

Coat

\u{1f9e5}: 'ğŸ¥'

Socks

\u{1f9e6}: 'ğŸ¦'

RedGiftEnvelope

\u{1f9e7}: 'ğŸ'

Firecracker

\u{1f9e8}: 'ğŸ¨'

JigsawPuzzlePiece

\u{1f9e9}: 'ğŸ©'

TestTube

\u{1f9ea}: 'ğŸª'

PetriDish

\u{1f9eb}: 'ğŸ«'

DnaDoubleHelix

\u{1f9ec}: 'ğŸ¬'

Compass

\u{1f9ed}: 'ğŸ­'

Abacus

\u{1f9ee}: 'ğŸ®'

FireExtinguisher

\u{1f9ef}: 'ğŸ¯'

Toolbox

\u{1f9f0}: 'ğŸ°'

Brick

\u{1f9f1}: 'ğŸ±'

Magnet

\u{1f9f2}: 'ğŸ²'

Luggage

\u{1f9f3}: 'ğŸ³'

LotionBottle

\u{1f9f4}: 'ğŸ´'

SpoolOfThread

\u{1f9f5}: 'ğŸµ'

BallOfYarn

\u{1f9f6}: 'ğŸ¶'

SafetyPin

\u{1f9f7}: 'ğŸ·'

TeddyBear

\u{1f9f8}: 'ğŸ¸'

Broom

\u{1f9f9}: 'ğŸ¹'

Basket

\u{1f9fa}: 'ğŸº'

RollOfPaper

\u{1f9fb}: 'ğŸ»'

BarOfSoap

\u{1f9fc}: 'ğŸ¼'

Sponge

\u{1f9fd}: 'ğŸ½'

Receipt

\u{1f9fe}: 'ğŸ¾'

Methods

impl SupplementalSymbolsandPictographs

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case