Enum unicode_types::syriac::Syriac

pub enum Syriac {
    EndOfParagraph,
    SupralinearFullStop,
    SublinearFullStop,
    SupralinearColon,
    SublinearColon,
    HorizontalColon,
    ColonSkewedLeft,
    ColonSkewedRight,
    SupralinearColonSkewedLeft,
    SublinearColonSkewedRight,
    Contraction,
    HarkleanObelus,
    HarkleanMetobelus,
    HarkleanAsteriscus,
    AbbreviationMark,
    LetterAlaph,
    LetterSuperscriptAlaph,
    LetterBeth,
    LetterGamal,
    LetterGamalGarshuni,
    LetterDalath,
    LetterDotlessDalathRish,
    LetterHe,
    LetterWaw,
    LetterZain,
    LetterHeth,
    LetterTeth,
    LetterTethGarshuni,
    LetterYudh,
    LetterYudhHe,
    LetterKaph,
    LetterLamadh,
    LetterMim,
    LetterNun,
    LetterSemkath,
    LetterFinalSemkath,
    LetterE,
    LetterPe,
    LetterReversedPe,
    LetterSadhe,
    LetterQaph,
    LetterRish,
    LetterShin,
    LetterTaw,
    LetterPersianBheth,
    LetterPersianGhamal,
    LetterPersianDhalath,
    PthahaAbove,
    PthahaBelow,
    PthahaDotted,
    ZqaphaAbove,
    ZqaphaBelow,
    ZqaphaDotted,
    RbasaAbove,
    RbasaBelow,
    DottedZlamaHorizontal,
    DottedZlamaAngular,
    HbasaAbove,
    HbasaBelow,
    HbasaDashEsasaDotted,
    EsasaAbove,
    EsasaBelow,
    Rwaha,
    FeminineDot,
    Qushshaya,
    Rukkakha,
    TwoVerticalDotsAbove,
    TwoVerticalDotsBelow,
    ThreeDotsAbove,
    ThreeDotsBelow,
    ObliqueLineAbove,
    ObliqueLineBelow,
    Music,
    Barrekh,
    LetterSogdianZhain,
    LetterSogdianKhaph,
}

An enum to represent all characters in the Syriac block.

Variants
EndOfParagraph

\u{700}: '܀'

SupralinearFullStop

\u{701}: '܁'

SublinearFullStop

\u{702}: '܂'

SupralinearColon

\u{703}: '܃'

SublinearColon

\u{704}: '܄'

HorizontalColon

\u{705}: '܅'

ColonSkewedLeft

\u{706}: '܆'

ColonSkewedRight

\u{707}: '܇'

SupralinearColonSkewedLeft

\u{708}: '܈'

SublinearColonSkewedRight

\u{709}: '܉'

Contraction

\u{70a}: '܊'

HarkleanObelus

\u{70b}: '܋'

HarkleanMetobelus

\u{70c}: '܌'

HarkleanAsteriscus

\u{70d}: '܍'

AbbreviationMark

\u{70f}: '܏'

LetterAlaph

\u{710}: 'ܐ'

LetterSuperscriptAlaph

\u{711}: 'ܑ'

LetterBeth

\u{712}: 'ܒ'

LetterGamal

\u{713}: 'ܓ'

LetterGamalGarshuni

\u{714}: 'ܔ'

LetterDalath

\u{715}: 'ܕ'

LetterDotlessDalathRish

\u{716}: 'ܖ'

LetterHe

\u{717}: 'ܗ'

LetterWaw

\u{718}: 'ܘ'

LetterZain

\u{719}: 'ܙ'

LetterHeth

\u{71a}: 'ܚ'

LetterTeth

\u{71b}: 'ܛ'

LetterTethGarshuni

\u{71c}: 'ܜ'

LetterYudh

\u{71d}: 'ܝ'

LetterYudhHe

\u{71e}: 'ܞ'

LetterKaph

\u{71f}: 'ܟ'

LetterLamadh

\u{720}: 'ܠ'

LetterMim

\u{721}: 'ܡ'

LetterNun

\u{722}: 'ܢ'

LetterSemkath

\u{723}: 'ܣ'

LetterFinalSemkath

\u{724}: 'ܤ'

LetterE

\u{725}: 'ܥ'

LetterPe

\u{726}: 'ܦ'

LetterReversedPe

\u{727}: 'ܧ'

LetterSadhe

\u{728}: 'ܨ'

LetterQaph

\u{729}: 'ܩ'

LetterRish

\u{72a}: 'ܪ'

LetterShin

\u{72b}: 'ܫ'

LetterTaw

\u{72c}: 'ܬ'

LetterPersianBheth

\u{72d}: 'ܭ'

LetterPersianGhamal

\u{72e}: 'ܮ'

LetterPersianDhalath

\u{72f}: 'ܯ'

PthahaAbove

\u{730}: 'ܰ'

PthahaBelow

\u{731}: 'ܱ'

PthahaDotted

\u{732}: 'ܲ'

ZqaphaAbove

\u{733}: 'ܳ'

ZqaphaBelow

\u{734}: 'ܴ'

ZqaphaDotted

\u{735}: 'ܵ'

RbasaAbove

\u{736}: 'ܶ'

RbasaBelow

\u{737}: 'ܷ'

DottedZlamaHorizontal

\u{738}: 'ܸ'

DottedZlamaAngular

\u{739}: 'ܹ'

HbasaAbove

\u{73a}: 'ܺ'

HbasaBelow

\u{73b}: 'ܻ'

HbasaDashEsasaDotted

\u{73c}: 'ܼ'

EsasaAbove

\u{73d}: 'ܽ'

EsasaBelow

\u{73e}: 'ܾ'

Rwaha

\u{73f}: 'ܿ'

FeminineDot

\u{740}: '݀'

Qushshaya

\u{741}: '݁'

Rukkakha

\u{742}: '݂'

TwoVerticalDotsAbove

\u{743}: '݃'

TwoVerticalDotsBelow

\u{744}: '݄'

ThreeDotsAbove

\u{745}: '݅'

ThreeDotsBelow

\u{746}: '݆'

ObliqueLineAbove

\u{747}: '݇'

ObliqueLineBelow

\u{748}: '݈'

Music

\u{749}: '݉'

Barrekh

\u{74a}: '݊'

LetterSogdianZhain

\u{74d}: 'ݍ'

LetterSogdianKhaph

\u{74e}: 'ݎ'

Methods

impl Syriac

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case