Enum unicode_types::tai_tham::TaiTham

pub enum TaiTham {
    LetterHighKa,
    LetterHighKha,
    LetterHighKxa,
    LetterLowKa,
    LetterLowKxa,
    LetterLowKha,
    LetterNga,
    LetterHighCa,
    LetterHighCha,
    LetterLowCa,
    LetterLowSa,
    LetterLowCha,
    LetterNya,
    LetterRata,
    LetterHighRatha,
    LetterDa,
    LetterLowRatha,
    LetterRana,
    LetterHighTa,
    LetterHighTha,
    LetterLowTa,
    LetterLowTha,
    LetterNa,
    LetterBa,
    LetterHighPa,
    LetterHighPha,
    LetterHighFa,
    LetterLowPa,
    LetterLowFa,
    LetterLowPha,
    LetterMa,
    LetterLowYa,
    LetterHighYa,
    LetterRa,
    LetterRue,
    LetterLa,
    LetterLue,
    LetterWa,
    LetterHighSha,
    LetterHighSsa,
    LetterHighSa,
    LetterHighHa,
    LetterLla,
    LetterA,
    LetterLowHa,
    LetterI,
    LetterIi,
    LetterU,
    LetterUu,
    LetterEe,
    LetterOo,
    LetterLae,
    LetterGreatSa,
    ConsonantSignMedialRa,
    ConsonantSignMedialLa,
    ConsonantSignLaTangLai,
    SignMaiKangLai,
    ConsonantSignFinalNga,
    ConsonantSignLowPa,
    ConsonantSignHighRathaOrLowPa,
    ConsonantSignMa,
    ConsonantSignBa,
    ConsonantSignSa,
    SignSakot,
    VowelSignA,
    VowelSignMaiSat,
    VowelSignAa,
    VowelSignTallAa,
    VowelSignI,
    VowelSignIi,
    VowelSignUe,
    VowelSignUue,
    VowelSignU,
    VowelSignUu,
    VowelSignO,
    VowelSignOaBelow,
    VowelSignOy,
    VowelSignE,
    VowelSignAe,
    VowelSignOo,
    VowelSignAi,
    VowelSignThamAi,
    VowelSignOaAbove,
    SignMaiKang,
    SignToneDash1,
    SignToneDash2,
    SignKhuenToneDash3,
    SignKhuenToneDash4,
    SignKhuenToneDash5,
    SignRaHaam,
    SignMaiSam,
    SignKhuenDashLueKaran,
    CombiningCryptogrammicDot,
    HoraDigitZero,
    HoraDigitOne,
    HoraDigitTwo,
    HoraDigitThree,
    HoraDigitFour,
    HoraDigitFive,
    HoraDigitSix,
    HoraDigitSeven,
    HoraDigitEight,
    HoraDigitNine,
    ThamDigitZero,
    ThamDigitOne,
    ThamDigitTwo,
    ThamDigitThree,
    ThamDigitFour,
    ThamDigitFive,
    ThamDigitSix,
    ThamDigitSeven,
    ThamDigitEight,
    ThamDigitNine,
    SignWiang,
    SignWiangwaak,
    SignSawan,
    SignKeow,
    SignHoy,
    SignDokmai,
    SignReversedRotatedRana,
    SignMaiYamok,
    SignKaan,
    SignKaankuu,
    SignSatkaan,
    SignSatkaankuu,
    SignHang,
    SignCaang,
}

An enum to represent all characters in the TaiTham block.

Variants
LetterHighKa

\u{1a20}: 'ᨠ'

LetterHighKha

\u{1a21}: 'ᨡ'

LetterHighKxa

\u{1a22}: 'ᨢ'

LetterLowKa

\u{1a23}: 'ᨣ'

LetterLowKxa

\u{1a24}: 'ᨤ'

LetterLowKha

\u{1a25}: 'ᨥ'

LetterNga

\u{1a26}: 'ᨦ'

LetterHighCa

\u{1a27}: 'ᨧ'

LetterHighCha

\u{1a28}: 'ᨨ'

LetterLowCa

\u{1a29}: 'ᨩ'

LetterLowSa

\u{1a2a}: 'ᨪ'

LetterLowCha

\u{1a2b}: 'ᨫ'

LetterNya

\u{1a2c}: 'ᨬ'

LetterRata

\u{1a2d}: 'ᨭ'

LetterHighRatha

\u{1a2e}: 'ᨮ'

LetterDa

\u{1a2f}: 'ᨯ'

LetterLowRatha

\u{1a30}: 'ᨰ'

LetterRana

\u{1a31}: 'ᨱ'

LetterHighTa

\u{1a32}: 'ᨲ'

LetterHighTha

\u{1a33}: 'ᨳ'

LetterLowTa

\u{1a34}: 'ᨴ'

LetterLowTha

\u{1a35}: 'ᨵ'

LetterNa

\u{1a36}: 'ᨶ'

LetterBa

\u{1a37}: 'ᨷ'

LetterHighPa

\u{1a38}: 'ᨸ'

LetterHighPha

\u{1a39}: 'ᨹ'

LetterHighFa

\u{1a3a}: 'ᨺ'

LetterLowPa

\u{1a3b}: 'ᨻ'

LetterLowFa

\u{1a3c}: 'ᨼ'

LetterLowPha

\u{1a3d}: 'ᨽ'

LetterMa

\u{1a3e}: 'ᨾ'

LetterLowYa

\u{1a3f}: 'ᨿ'

LetterHighYa

\u{1a40}: 'ᩀ'

LetterRa

\u{1a41}: 'ᩁ'

LetterRue

\u{1a42}: 'ᩂ'

LetterLa

\u{1a43}: 'ᩃ'

LetterLue

\u{1a44}: 'ᩄ'

LetterWa

\u{1a45}: 'ᩅ'

LetterHighSha

\u{1a46}: 'ᩆ'

LetterHighSsa

\u{1a47}: 'ᩇ'

LetterHighSa

\u{1a48}: 'ᩈ'

LetterHighHa

\u{1a49}: 'ᩉ'

LetterLla

\u{1a4a}: 'ᩊ'

LetterA

\u{1a4b}: 'ᩋ'

LetterLowHa

\u{1a4c}: 'ᩌ'

LetterI

\u{1a4d}: 'ᩍ'

LetterIi

\u{1a4e}: 'ᩎ'

LetterU

\u{1a4f}: 'ᩏ'

LetterUu

\u{1a50}: 'ᩐ'

LetterEe

\u{1a51}: 'ᩑ'

LetterOo

\u{1a52}: 'ᩒ'

LetterLae

\u{1a53}: 'ᩓ'

LetterGreatSa

\u{1a54}: 'ᩔ'

ConsonantSignMedialRa

\u{1a55}: 'ᩕ'

ConsonantSignMedialLa

\u{1a56}: 'ᩖ'

ConsonantSignLaTangLai

\u{1a57}: 'ᩗ'

SignMaiKangLai

\u{1a58}: 'ᩘ'

ConsonantSignFinalNga

\u{1a59}: 'ᩙ'

ConsonantSignLowPa

\u{1a5a}: 'ᩚ'

ConsonantSignHighRathaOrLowPa

\u{1a5b}: 'ᩛ'

ConsonantSignMa

\u{1a5c}: 'ᩜ'

ConsonantSignBa

\u{1a5d}: 'ᩝ'

ConsonantSignSa

\u{1a5e}: 'ᩞ'

SignSakot

\u{1a60}: '᩠'

VowelSignA

\u{1a61}: 'ᩡ'

VowelSignMaiSat

\u{1a62}: 'ᩢ'

VowelSignAa

\u{1a63}: 'ᩣ'

VowelSignTallAa

\u{1a64}: 'ᩤ'

VowelSignI

\u{1a65}: 'ᩥ'

VowelSignIi

\u{1a66}: 'ᩦ'

VowelSignUe

\u{1a67}: 'ᩧ'

VowelSignUue

\u{1a68}: 'ᩨ'

VowelSignU

\u{1a69}: 'ᩩ'

VowelSignUu

\u{1a6a}: 'ᩪ'

VowelSignO

\u{1a6b}: 'ᩫ'

VowelSignOaBelow

\u{1a6c}: 'ᩬ'

VowelSignOy

\u{1a6d}: 'ᩭ'

VowelSignE

\u{1a6e}: 'ᩮ'

VowelSignAe

\u{1a6f}: 'ᩯ'

VowelSignOo

\u{1a70}: 'ᩰ'

VowelSignAi

\u{1a71}: 'ᩱ'

VowelSignThamAi

\u{1a72}: 'ᩲ'

VowelSignOaAbove

\u{1a73}: 'ᩳ'

SignMaiKang

\u{1a74}: 'ᩴ'

SignToneDash1

\u{1a75}: '᩵'

SignToneDash2

\u{1a76}: '᩶'

SignKhuenToneDash3

\u{1a77}: '᩷'

SignKhuenToneDash4

\u{1a78}: '᩸'

SignKhuenToneDash5

\u{1a79}: '᩹'

SignRaHaam

\u{1a7a}: '᩺'

SignMaiSam

\u{1a7b}: '᩻'

SignKhuenDashLueKaran

\u{1a7c}: '᩼'

CombiningCryptogrammicDot

\u{1a7f}: '᩿'

HoraDigitZero

\u{1a80}: '᪀'

HoraDigitOne

\u{1a81}: '᪁'

HoraDigitTwo

\u{1a82}: '᪂'

HoraDigitThree

\u{1a83}: '᪃'

HoraDigitFour

\u{1a84}: '᪄'

HoraDigitFive

\u{1a85}: '᪅'

HoraDigitSix

\u{1a86}: '᪆'

HoraDigitSeven

\u{1a87}: '᪇'

HoraDigitEight

\u{1a88}: '᪈'

HoraDigitNine

\u{1a89}: '᪉'

ThamDigitZero

\u{1a90}: '᪐'

ThamDigitOne

\u{1a91}: '᪑'

ThamDigitTwo

\u{1a92}: '᪒'

ThamDigitThree

\u{1a93}: '᪓'

ThamDigitFour

\u{1a94}: '᪔'

ThamDigitFive

\u{1a95}: '᪕'

ThamDigitSix

\u{1a96}: '᪖'

ThamDigitSeven

\u{1a97}: '᪗'

ThamDigitEight

\u{1a98}: '᪘'

ThamDigitNine

\u{1a99}: '᪙'

SignWiang

\u{1aa0}: '᪠'

SignWiangwaak

\u{1aa1}: '᪡'

SignSawan

\u{1aa2}: '᪢'

SignKeow

\u{1aa3}: '᪣'

SignHoy

\u{1aa4}: '᪤'

SignDokmai

\u{1aa5}: '᪥'

SignReversedRotatedRana

\u{1aa6}: '᪦'

SignMaiYamok

\u{1aa7}: 'ᪧ'

SignKaan

\u{1aa8}: '᪨'

SignKaankuu

\u{1aa9}: '᪩'

SignSatkaan

\u{1aaa}: '᪪'

SignSatkaankuu

\u{1aab}: '᪫'

SignHang

\u{1aac}: '᪬'

SignCaang

\u{1aad}: '᪭'

Methods

impl TaiTham

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case