Enum unicode_types::tai_viet::TaiViet

pub enum TaiViet {
    LetterLowKo,
    LetterHighKo,
    LetterLowKho,
    LetterHighKho,
    LetterLowKhho,
    LetterHighKhho,
    LetterLowGo,
    LetterHighGo,
    LetterLowNgo,
    LetterHighNgo,
    LetterLowCo,
    LetterHighCo,
    LetterLowCho,
    LetterHighCho,
    LetterLowSo,
    LetterHighSo,
    LetterLowNyo,
    LetterHighNyo,
    LetterLowDo,
    LetterHighDo,
    LetterLowTo,
    LetterHighTo,
    LetterLowTho,
    LetterHighTho,
    LetterLowNo,
    LetterHighNo,
    LetterLowBo,
    LetterHighBo,
    LetterLowPo,
    LetterHighPo,
    LetterLowPho,
    LetterHighPho,
    LetterLowFo,
    LetterHighFo,
    LetterLowMo,
    LetterHighMo,
    LetterLowYo,
    LetterHighYo,
    LetterLowRo,
    LetterHighRo,
    LetterLowLo,
    LetterHighLo,
    LetterLowVo,
    LetterHighVo,
    LetterLowHo,
    LetterHighHo,
    LetterLowO,
    LetterHighO,
    MaiKang,
    VowelAa,
    VowelI,
    VowelUe,
    VowelU,
    VowelE,
    VowelO,
    MaiKhit,
    VowelIa,
    VowelUea,
    VowelUa,
    VowelAue,
    VowelAy,
    VowelAn,
    VowelAm,
    ToneMaiEk,
    ToneMaiNueng,
    ToneMaiTho,
    ToneMaiSong,
    SymbolKon,
    SymbolNueng,
    SymbolSam,
    SymbolHoHoi,
}

An enum to represent all characters in the TaiViet block.

Variants
LetterLowKo

\u{aa80}: 'ꪀ'

LetterHighKo

\u{aa81}: 'ꪁ'

LetterLowKho

\u{aa82}: 'ꪂ'

LetterHighKho

\u{aa83}: 'ꪃ'

LetterLowKhho

\u{aa84}: 'ꪄ'

LetterHighKhho

\u{aa85}: 'ꪅ'

LetterLowGo

\u{aa86}: 'ꪆ'

LetterHighGo

\u{aa87}: 'ꪇ'

LetterLowNgo

\u{aa88}: 'ꪈ'

LetterHighNgo

\u{aa89}: 'ꪉ'

LetterLowCo

\u{aa8a}: 'ꪊ'

LetterHighCo

\u{aa8b}: 'ꪋ'

LetterLowCho

\u{aa8c}: 'ꪌ'

LetterHighCho

\u{aa8d}: 'ꪍ'

LetterLowSo

\u{aa8e}: 'ꪎ'

LetterHighSo

\u{aa8f}: 'ꪏ'

LetterLowNyo

\u{aa90}: 'ꪐ'

LetterHighNyo

\u{aa91}: 'ꪑ'

LetterLowDo

\u{aa92}: 'ꪒ'

LetterHighDo

\u{aa93}: 'ꪓ'

LetterLowTo

\u{aa94}: 'ꪔ'

LetterHighTo

\u{aa95}: 'ꪕ'

LetterLowTho

\u{aa96}: 'ꪖ'

LetterHighTho

\u{aa97}: 'ꪗ'

LetterLowNo

\u{aa98}: 'ꪘ'

LetterHighNo

\u{aa99}: 'ꪙ'

LetterLowBo

\u{aa9a}: 'ꪚ'

LetterHighBo

\u{aa9b}: 'ꪛ'

LetterLowPo

\u{aa9c}: 'ꪜ'

LetterHighPo

\u{aa9d}: 'ꪝ'

LetterLowPho

\u{aa9e}: 'ꪞ'

LetterHighPho

\u{aa9f}: 'ꪟ'

LetterLowFo

\u{aaa0}: 'ꪠ'

LetterHighFo

\u{aaa1}: 'ꪡ'

LetterLowMo

\u{aaa2}: 'ꪢ'

LetterHighMo

\u{aaa3}: 'ꪣ'

LetterLowYo

\u{aaa4}: 'ꪤ'

LetterHighYo

\u{aaa5}: 'ꪥ'

LetterLowRo

\u{aaa6}: 'ꪦ'

LetterHighRo

\u{aaa7}: 'ꪧ'

LetterLowLo

\u{aaa8}: 'ꪨ'

LetterHighLo

\u{aaa9}: 'ꪩ'

LetterLowVo

\u{aaaa}: 'ꪪ'

LetterHighVo

\u{aaab}: 'ꪫ'

LetterLowHo

\u{aaac}: 'ꪬ'

LetterHighHo

\u{aaad}: 'ꪭ'

LetterLowO

\u{aaae}: 'ꪮ'

LetterHighO

\u{aaaf}: 'ꪯ'

MaiKang

\u{aab0}: 'ꪰ'

VowelAa

\u{aab1}: 'ꪱ'

VowelI

\u{aab2}: 'ꪲ'

VowelUe

\u{aab3}: 'ꪳ'

VowelU

\u{aab4}: 'ꪴ'

VowelE

\u{aab5}: 'ꪵ'

VowelO

\u{aab6}: 'ꪶ'

MaiKhit

\u{aab7}: 'ꪷ'

VowelIa

\u{aab8}: 'ꪸ'

VowelUea

\u{aab9}: 'ꪹ'

VowelUa

\u{aaba}: 'ꪺ'

VowelAue

\u{aabb}: 'ꪻ'

VowelAy

\u{aabc}: 'ꪼ'

VowelAn

\u{aabd}: 'ꪽ'

VowelAm

\u{aabe}: 'ꪾ'

ToneMaiEk

\u{aabf}: '꪿'

ToneMaiNueng

\u{aac0}: 'ꫀ'

ToneMaiTho

\u{aac1}: '꫁'

ToneMaiSong

\u{aac2}: 'ꫂ'

SymbolKon

\u{aadb}: 'ꫛ'

SymbolNueng

\u{aadc}: 'ꫜ'

SymbolSam

\u{aadd}: 'ꫝ'

SymbolHoHoi

\u{aade}: '꫞'

Methods

impl TaiViet

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case