Enum unicode_types::tangut_components::TangutComponents

pub enum TangutComponents {
    TangutComponentDash001,
    TangutComponentDash002,
    TangutComponentDash003,
    TangutComponentDash004,
    TangutComponentDash005,
    TangutComponentDash006,
    TangutComponentDash007,
    TangutComponentDash008,
    TangutComponentDash009,
    TangutComponentDash010,
    TangutComponentDash011,
    TangutComponentDash012,
    TangutComponentDash013,
    TangutComponentDash014,
    TangutComponentDash015,
    TangutComponentDash016,
    TangutComponentDash017,
    TangutComponentDash018,
    TangutComponentDash019,
    TangutComponentDash020,
    TangutComponentDash021,
    TangutComponentDash022,
    TangutComponentDash023,
    TangutComponentDash024,
    TangutComponentDash025,
    TangutComponentDash026,
    TangutComponentDash027,
    TangutComponentDash028,
    TangutComponentDash029,
    TangutComponentDash030,
    TangutComponentDash031,
    TangutComponentDash032,
    TangutComponentDash033,
    TangutComponentDash034,
    TangutComponentDash035,
    TangutComponentDash036,
    TangutComponentDash037,
    TangutComponentDash038,
    TangutComponentDash039,
    TangutComponentDash040,
    TangutComponentDash041,
    TangutComponentDash042,
    TangutComponentDash043,
    TangutComponentDash044,
    TangutComponentDash045,
    TangutComponentDash046,
    TangutComponentDash047,
    TangutComponentDash048,
    TangutComponentDash049,
    TangutComponentDash050,
    TangutComponentDash051,
    TangutComponentDash052,
    TangutComponentDash053,
    TangutComponentDash054,
    TangutComponentDash055,
    TangutComponentDash056,
    TangutComponentDash057,
    TangutComponentDash058,
    TangutComponentDash059,
    TangutComponentDash060,
    TangutComponentDash061,
    TangutComponentDash062,
    TangutComponentDash063,
    TangutComponentDash064,
    TangutComponentDash065,
    TangutComponentDash066,
    TangutComponentDash067,
    TangutComponentDash068,
    TangutComponentDash069,
    TangutComponentDash070,
    TangutComponentDash071,
    TangutComponentDash072,
    TangutComponentDash073,
    TangutComponentDash074,
    TangutComponentDash075,
    TangutComponentDash076,
    TangutComponentDash077,
    TangutComponentDash078,
    TangutComponentDash079,
    TangutComponentDash080,
    TangutComponentDash081,
    TangutComponentDash082,
    TangutComponentDash083,
    TangutComponentDash084,
    TangutComponentDash085,
    TangutComponentDash086,
    TangutComponentDash087,
    TangutComponentDash088,
    TangutComponentDash089,
    TangutComponentDash090,
    TangutComponentDash091,
    TangutComponentDash092,
    TangutComponentDash093,
    TangutComponentDash094,
    TangutComponentDash095,
    TangutComponentDash096,
    TangutComponentDash097,
    TangutComponentDash098,
    TangutComponentDash099,
    TangutComponentDash100,
    TangutComponentDash101,
    TangutComponentDash102,
    TangutComponentDash103,
    TangutComponentDash104,
    TangutComponentDash105,
    TangutComponentDash106,
    TangutComponentDash107,
    TangutComponentDash108,
    TangutComponentDash109,
    TangutComponentDash110,
    TangutComponentDash111,
    TangutComponentDash112,
    TangutComponentDash113,
    TangutComponentDash114,
    TangutComponentDash115,
    TangutComponentDash116,
    TangutComponentDash117,
    TangutComponentDash118,
    TangutComponentDash119,
    TangutComponentDash120,
    TangutComponentDash121,
    TangutComponentDash122,
    TangutComponentDash123,
    TangutComponentDash124,
    TangutComponentDash125,
    TangutComponentDash126,
    TangutComponentDash127,
    TangutComponentDash128,
    TangutComponentDash129,
    TangutComponentDash130,
    TangutComponentDash131,
    TangutComponentDash132,
    TangutComponentDash133,
    TangutComponentDash134,
    TangutComponentDash135,
    TangutComponentDash136,
    TangutComponentDash137,
    TangutComponentDash138,
    TangutComponentDash139,
    TangutComponentDash140,
    TangutComponentDash141,
    TangutComponentDash142,
    TangutComponentDash143,
    TangutComponentDash144,
    TangutComponentDash145,
    TangutComponentDash146,
    TangutComponentDash147,
    TangutComponentDash148,
    TangutComponentDash149,
    TangutComponentDash150,
    TangutComponentDash151,
    TangutComponentDash152,
    TangutComponentDash153,
    TangutComponentDash154,
    TangutComponentDash155,
    TangutComponentDash156,
    TangutComponentDash157,
    TangutComponentDash158,
    TangutComponentDash159,
    TangutComponentDash160,
    TangutComponentDash161,
    TangutComponentDash162,
    TangutComponentDash163,
    TangutComponentDash164,
    TangutComponentDash165,
    TangutComponentDash166,
    TangutComponentDash167,
    TangutComponentDash168,
    TangutComponentDash169,
    TangutComponentDash170,
    TangutComponentDash171,
    TangutComponentDash172,
    TangutComponentDash173,
    TangutComponentDash174,
    TangutComponentDash175,
    TangutComponentDash176,
    TangutComponentDash177,
    TangutComponentDash178,
    TangutComponentDash179,
    TangutComponentDash180,
    TangutComponentDash181,
    TangutComponentDash182,
    TangutComponentDash183,
    TangutComponentDash184,
    TangutComponentDash185,
    TangutComponentDash186,
    TangutComponentDash187,
    TangutComponentDash188,
    TangutComponentDash189,
    TangutComponentDash190,
    TangutComponentDash191,
    TangutComponentDash192,
    TangutComponentDash193,
    TangutComponentDash194,
    TangutComponentDash195,
    TangutComponentDash196,
    TangutComponentDash197,
    TangutComponentDash198,
    TangutComponentDash199,
    TangutComponentDash200,
    TangutComponentDash201,
    TangutComponentDash202,
    TangutComponentDash203,
    TangutComponentDash204,
    TangutComponentDash205,
    TangutComponentDash206,
    TangutComponentDash207,
    TangutComponentDash208,
    TangutComponentDash209,
    TangutComponentDash210,
    TangutComponentDash211,
    TangutComponentDash212,
    TangutComponentDash213,
    TangutComponentDash214,
    TangutComponentDash215,
    TangutComponentDash216,
    TangutComponentDash217,
    TangutComponentDash218,
    TangutComponentDash219,
    TangutComponentDash220,
    TangutComponentDash221,
    TangutComponentDash222,
    TangutComponentDash223,
    TangutComponentDash224,
    TangutComponentDash225,
    TangutComponentDash226,
    TangutComponentDash227,
    TangutComponentDash228,
    TangutComponentDash229,
    TangutComponentDash230,
    TangutComponentDash231,
    TangutComponentDash232,
    TangutComponentDash233,
    TangutComponentDash234,
    TangutComponentDash235,
    TangutComponentDash236,
    TangutComponentDash237,
    TangutComponentDash238,
    TangutComponentDash239,
    TangutComponentDash240,
    TangutComponentDash241,
    TangutComponentDash242,
    TangutComponentDash243,
    TangutComponentDash244,
    TangutComponentDash245,
    TangutComponentDash246,
    TangutComponentDash247,
    TangutComponentDash248,
    TangutComponentDash249,
    TangutComponentDash250,
    TangutComponentDash251,
    TangutComponentDash252,
    TangutComponentDash253,
    TangutComponentDash254,
    TangutComponentDash255,
    TangutComponentDash256,
    TangutComponentDash257,
    TangutComponentDash258,
    TangutComponentDash259,
    TangutComponentDash260,
    TangutComponentDash261,
    TangutComponentDash262,
    TangutComponentDash263,
    TangutComponentDash264,
    TangutComponentDash265,
    TangutComponentDash266,
    TangutComponentDash267,
    TangutComponentDash268,
    TangutComponentDash269,
    TangutComponentDash270,
    TangutComponentDash271,
    TangutComponentDash272,
    TangutComponentDash273,
    TangutComponentDash274,
    TangutComponentDash275,
    TangutComponentDash276,
    TangutComponentDash277,
    TangutComponentDash278,
    TangutComponentDash279,
    TangutComponentDash280,
    TangutComponentDash281,
    TangutComponentDash282,
    TangutComponentDash283,
    TangutComponentDash284,
    TangutComponentDash285,
    TangutComponentDash286,
    TangutComponentDash287,
    TangutComponentDash288,
    TangutComponentDash289,
    TangutComponentDash290,
    TangutComponentDash291,
    TangutComponentDash292,
    TangutComponentDash293,
    TangutComponentDash294,
    TangutComponentDash295,
    TangutComponentDash296,
    TangutComponentDash297,
    TangutComponentDash298,
    TangutComponentDash299,
    TangutComponentDash300,
    TangutComponentDash301,
    TangutComponentDash302,
    TangutComponentDash303,
    TangutComponentDash304,
    TangutComponentDash305,
    TangutComponentDash306,
    TangutComponentDash307,
    TangutComponentDash308,
    TangutComponentDash309,
    TangutComponentDash310,
    TangutComponentDash311,
    TangutComponentDash312,
    TangutComponentDash313,
    TangutComponentDash314,
    TangutComponentDash315,
    TangutComponentDash316,
    TangutComponentDash317,
    TangutComponentDash318,
    TangutComponentDash319,
    TangutComponentDash320,
    TangutComponentDash321,
    TangutComponentDash322,
    TangutComponentDash323,
    TangutComponentDash324,
    TangutComponentDash325,
    TangutComponentDash326,
    TangutComponentDash327,
    TangutComponentDash328,
    TangutComponentDash329,
    TangutComponentDash330,
    TangutComponentDash331,
    TangutComponentDash332,
    TangutComponentDash333,
    TangutComponentDash334,
    TangutComponentDash335,
    TangutComponentDash336,
    TangutComponentDash337,
    TangutComponentDash338,
    TangutComponentDash339,
    TangutComponentDash340,
    TangutComponentDash341,
    TangutComponentDash342,
    TangutComponentDash343,
    TangutComponentDash344,
    TangutComponentDash345,
    TangutComponentDash346,
    TangutComponentDash347,
    TangutComponentDash348,
    TangutComponentDash349,
    TangutComponentDash350,
    TangutComponentDash351,
    TangutComponentDash352,
    TangutComponentDash353,
    TangutComponentDash354,
    TangutComponentDash355,
    TangutComponentDash356,
    TangutComponentDash357,
    TangutComponentDash358,
    TangutComponentDash359,
    TangutComponentDash360,
    TangutComponentDash361,
    TangutComponentDash362,
    TangutComponentDash363,
    TangutComponentDash364,
    TangutComponentDash365,
    TangutComponentDash366,
    TangutComponentDash367,
    TangutComponentDash368,
    TangutComponentDash369,
    TangutComponentDash370,
    TangutComponentDash371,
    TangutComponentDash372,
    TangutComponentDash373,
    TangutComponentDash374,
    TangutComponentDash375,
    TangutComponentDash376,
    TangutComponentDash377,
    TangutComponentDash378,
    TangutComponentDash379,
    TangutComponentDash380,
    TangutComponentDash381,
    TangutComponentDash382,
    TangutComponentDash383,
    TangutComponentDash384,
    TangutComponentDash385,
    TangutComponentDash386,
    TangutComponentDash387,
    TangutComponentDash388,
    TangutComponentDash389,
    TangutComponentDash390,
    TangutComponentDash391,
    TangutComponentDash392,
    TangutComponentDash393,
    TangutComponentDash394,
    TangutComponentDash395,
    TangutComponentDash396,
    TangutComponentDash397,
    TangutComponentDash398,
    TangutComponentDash399,
    TangutComponentDash400,
    TangutComponentDash401,
    TangutComponentDash402,
    TangutComponentDash403,
    TangutComponentDash404,
    TangutComponentDash405,
    TangutComponentDash406,
    TangutComponentDash407,
    TangutComponentDash408,
    TangutComponentDash409,
    TangutComponentDash410,
    TangutComponentDash411,
    TangutComponentDash412,
    TangutComponentDash413,
    TangutComponentDash414,
    TangutComponentDash415,
    TangutComponentDash416,
    TangutComponentDash417,
    TangutComponentDash418,
    TangutComponentDash419,
    TangutComponentDash420,
    TangutComponentDash421,
    TangutComponentDash422,
    TangutComponentDash423,
    TangutComponentDash424,
    TangutComponentDash425,
    TangutComponentDash426,
    TangutComponentDash427,
    TangutComponentDash428,
    TangutComponentDash429,
    TangutComponentDash430,
    TangutComponentDash431,
    TangutComponentDash432,
    TangutComponentDash433,
    TangutComponentDash434,
    TangutComponentDash435,
    TangutComponentDash436,
    TangutComponentDash437,
    TangutComponentDash438,
    TangutComponentDash439,
    TangutComponentDash440,
    TangutComponentDash441,
    TangutComponentDash442,
    TangutComponentDash443,
    TangutComponentDash444,
    TangutComponentDash445,
    TangutComponentDash446,
    TangutComponentDash447,
    TangutComponentDash448,
    TangutComponentDash449,
    TangutComponentDash450,
    TangutComponentDash451,
    TangutComponentDash452,
    TangutComponentDash453,
    TangutComponentDash454,
    TangutComponentDash455,
    TangutComponentDash456,
    TangutComponentDash457,
    TangutComponentDash458,
    TangutComponentDash459,
    TangutComponentDash460,
    TangutComponentDash461,
    TangutComponentDash462,
    TangutComponentDash463,
    TangutComponentDash464,
    TangutComponentDash465,
    TangutComponentDash466,
    TangutComponentDash467,
    TangutComponentDash468,
    TangutComponentDash469,
    TangutComponentDash470,
    TangutComponentDash471,
    TangutComponentDash472,
    TangutComponentDash473,
    TangutComponentDash474,
    TangutComponentDash475,
    TangutComponentDash476,
    TangutComponentDash477,
    TangutComponentDash478,
    TangutComponentDash479,
    TangutComponentDash480,
    TangutComponentDash481,
    TangutComponentDash482,
    TangutComponentDash483,
    TangutComponentDash484,
    TangutComponentDash485,
    TangutComponentDash486,
    TangutComponentDash487,
    TangutComponentDash488,
    TangutComponentDash489,
    TangutComponentDash490,
    TangutComponentDash491,
    TangutComponentDash492,
    TangutComponentDash493,
    TangutComponentDash494,
    TangutComponentDash495,
    TangutComponentDash496,
    TangutComponentDash497,
    TangutComponentDash498,
    TangutComponentDash499,
    TangutComponentDash500,
    TangutComponentDash501,
    TangutComponentDash502,
    TangutComponentDash503,
    TangutComponentDash504,
    TangutComponentDash505,
    TangutComponentDash506,
    TangutComponentDash507,
    TangutComponentDash508,
    TangutComponentDash509,
    TangutComponentDash510,
    TangutComponentDash511,
    TangutComponentDash512,
    TangutComponentDash513,
    TangutComponentDash514,
    TangutComponentDash515,
    TangutComponentDash516,
    TangutComponentDash517,
    TangutComponentDash518,
    TangutComponentDash519,
    TangutComponentDash520,
    TangutComponentDash521,
    TangutComponentDash522,
    TangutComponentDash523,
    TangutComponentDash524,
    TangutComponentDash525,
    TangutComponentDash526,
    TangutComponentDash527,
    TangutComponentDash528,
    TangutComponentDash529,
    TangutComponentDash530,
    TangutComponentDash531,
    TangutComponentDash532,
    TangutComponentDash533,
    TangutComponentDash534,
    TangutComponentDash535,
    TangutComponentDash536,
    TangutComponentDash537,
    TangutComponentDash538,
    TangutComponentDash539,
    TangutComponentDash540,
    TangutComponentDash541,
    TangutComponentDash542,
    TangutComponentDash543,
    TangutComponentDash544,
    TangutComponentDash545,
    TangutComponentDash546,
    TangutComponentDash547,
    TangutComponentDash548,
    TangutComponentDash549,
    TangutComponentDash550,
    TangutComponentDash551,
    TangutComponentDash552,
    TangutComponentDash553,
    TangutComponentDash554,
    TangutComponentDash555,
    TangutComponentDash556,
    TangutComponentDash557,
    TangutComponentDash558,
    TangutComponentDash559,
    TangutComponentDash560,
    TangutComponentDash561,
    TangutComponentDash562,
    TangutComponentDash563,
    TangutComponentDash564,
    TangutComponentDash565,
    TangutComponentDash566,
    TangutComponentDash567,
    TangutComponentDash568,
    TangutComponentDash569,
    TangutComponentDash570,
    TangutComponentDash571,
    TangutComponentDash572,
    TangutComponentDash573,
    TangutComponentDash574,
    TangutComponentDash575,
    TangutComponentDash576,
    TangutComponentDash577,
    TangutComponentDash578,
    TangutComponentDash579,
    TangutComponentDash580,
    TangutComponentDash581,
    TangutComponentDash582,
    TangutComponentDash583,
    TangutComponentDash584,
    TangutComponentDash585,
    TangutComponentDash586,
    TangutComponentDash587,
    TangutComponentDash588,
    TangutComponentDash589,
    TangutComponentDash590,
    TangutComponentDash591,
    TangutComponentDash592,
    TangutComponentDash593,
    TangutComponentDash594,
    TangutComponentDash595,
    TangutComponentDash596,
    TangutComponentDash597,
    TangutComponentDash598,
    TangutComponentDash599,
    TangutComponentDash600,
    TangutComponentDash601,
    TangutComponentDash602,
    TangutComponentDash603,
    TangutComponentDash604,
    TangutComponentDash605,
    TangutComponentDash606,
    TangutComponentDash607,
    TangutComponentDash608,
    TangutComponentDash609,
    TangutComponentDash610,
    TangutComponentDash611,
    TangutComponentDash612,
    TangutComponentDash613,
    TangutComponentDash614,
    TangutComponentDash615,
    TangutComponentDash616,
    TangutComponentDash617,
    TangutComponentDash618,
    TangutComponentDash619,
    TangutComponentDash620,
    TangutComponentDash621,
    TangutComponentDash622,
    TangutComponentDash623,
    TangutComponentDash624,
    TangutComponentDash625,
    TangutComponentDash626,
    TangutComponentDash627,
    TangutComponentDash628,
    TangutComponentDash629,
    TangutComponentDash630,
    TangutComponentDash631,
    TangutComponentDash632,
    TangutComponentDash633,
    TangutComponentDash634,
    TangutComponentDash635,
    TangutComponentDash636,
    TangutComponentDash637,
    TangutComponentDash638,
    TangutComponentDash639,
    TangutComponentDash640,
    TangutComponentDash641,
    TangutComponentDash642,
    TangutComponentDash643,
    TangutComponentDash644,
    TangutComponentDash645,
    TangutComponentDash646,
    TangutComponentDash647,
    TangutComponentDash648,
    TangutComponentDash649,
    TangutComponentDash650,
    TangutComponentDash651,
    TangutComponentDash652,
    TangutComponentDash653,
    TangutComponentDash654,
    TangutComponentDash655,
    TangutComponentDash656,
    TangutComponentDash657,
    TangutComponentDash658,
    TangutComponentDash659,
    TangutComponentDash660,
    TangutComponentDash661,
    TangutComponentDash662,
    TangutComponentDash663,
    TangutComponentDash664,
    TangutComponentDash665,
    TangutComponentDash666,
    TangutComponentDash667,
    TangutComponentDash668,
    TangutComponentDash669,
    TangutComponentDash670,
    TangutComponentDash671,
    TangutComponentDash672,
    TangutComponentDash673,
    TangutComponentDash674,
    TangutComponentDash675,
    TangutComponentDash676,
    TangutComponentDash677,
    TangutComponentDash678,
    TangutComponentDash679,
    TangutComponentDash680,
    TangutComponentDash681,
    TangutComponentDash682,
    TangutComponentDash683,
    TangutComponentDash684,
    TangutComponentDash685,
    TangutComponentDash686,
    TangutComponentDash687,
    TangutComponentDash688,
    TangutComponentDash689,
    TangutComponentDash690,
    TangutComponentDash691,
    TangutComponentDash692,
    TangutComponentDash693,
    TangutComponentDash694,
    TangutComponentDash695,
    TangutComponentDash696,
    TangutComponentDash697,
    TangutComponentDash698,
    TangutComponentDash699,
    TangutComponentDash700,
    TangutComponentDash701,
    TangutComponentDash702,
    TangutComponentDash703,
    TangutComponentDash704,
    TangutComponentDash705,
    TangutComponentDash706,
    TangutComponentDash707,
    TangutComponentDash708,
    TangutComponentDash709,
    TangutComponentDash710,
    TangutComponentDash711,
    TangutComponentDash712,
    TangutComponentDash713,
    TangutComponentDash714,
    TangutComponentDash715,
    TangutComponentDash716,
    TangutComponentDash717,
    TangutComponentDash718,
    TangutComponentDash719,
    TangutComponentDash720,
    TangutComponentDash721,
    TangutComponentDash722,
    TangutComponentDash723,
    TangutComponentDash724,
    TangutComponentDash725,
    TangutComponentDash726,
    TangutComponentDash727,
    TangutComponentDash728,
    TangutComponentDash729,
    TangutComponentDash730,
    TangutComponentDash731,
    TangutComponentDash732,
    TangutComponentDash733,
    TangutComponentDash734,
    TangutComponentDash735,
    TangutComponentDash736,
    TangutComponentDash737,
    TangutComponentDash738,
    TangutComponentDash739,
    TangutComponentDash740,
    TangutComponentDash741,
    TangutComponentDash742,
    TangutComponentDash743,
    TangutComponentDash744,
    TangutComponentDash745,
    TangutComponentDash746,
    TangutComponentDash747,
    TangutComponentDash748,
    TangutComponentDash749,
    TangutComponentDash750,
    TangutComponentDash751,
    TangutComponentDash752,
    TangutComponentDash753,
    TangutComponentDash754,
    TangutComponentDash755,
}

An enum to represent all characters in the TangutComponents block.

Variants
TangutComponentDash001

\u{18800}: ''

TangutComponentDash002

\u{18801}: ''

TangutComponentDash003

\u{18802}: ''

TangutComponentDash004

\u{18803}: ''

TangutComponentDash005

\u{18804}: ''

TangutComponentDash006

\u{18805}: ''

TangutComponentDash007

\u{18806}: ''

TangutComponentDash008

\u{18807}: ''

TangutComponentDash009

\u{18808}: ''

TangutComponentDash010

\u{18809}: ''

TangutComponentDash011

\u{1880a}: ''

TangutComponentDash012

\u{1880b}: ''

TangutComponentDash013

\u{1880c}: ''

TangutComponentDash014

\u{1880d}: ''

TangutComponentDash015

\u{1880e}: ''

TangutComponentDash016

\u{1880f}: ''

TangutComponentDash017

\u{18810}: ''

TangutComponentDash018

\u{18811}: ''

TangutComponentDash019

\u{18812}: ''

TangutComponentDash020

\u{18813}: ''

TangutComponentDash021

\u{18814}: ''

TangutComponentDash022

\u{18815}: ''

TangutComponentDash023

\u{18816}: ''

TangutComponentDash024

\u{18817}: ''

TangutComponentDash025

\u{18818}: ''

TangutComponentDash026

\u{18819}: ''

TangutComponentDash027

\u{1881a}: ''

TangutComponentDash028

\u{1881b}: ''

TangutComponentDash029

\u{1881c}: ''

TangutComponentDash030

\u{1881d}: ''

TangutComponentDash031

\u{1881e}: ''

TangutComponentDash032

\u{1881f}: ''

TangutComponentDash033

\u{18820}: ''

TangutComponentDash034

\u{18821}: ''

TangutComponentDash035

\u{18822}: ''

TangutComponentDash036

\u{18823}: ''

TangutComponentDash037

\u{18824}: ''

TangutComponentDash038

\u{18825}: ''

TangutComponentDash039

\u{18826}: ''

TangutComponentDash040

\u{18827}: ''

TangutComponentDash041

\u{18828}: ''

TangutComponentDash042

\u{18829}: ''

TangutComponentDash043

\u{1882a}: ''

TangutComponentDash044

\u{1882b}: ''

TangutComponentDash045

\u{1882c}: ''

TangutComponentDash046

\u{1882d}: ''

TangutComponentDash047

\u{1882e}: ''

TangutComponentDash048

\u{1882f}: ''

TangutComponentDash049

\u{18830}: ''

TangutComponentDash050

\u{18831}: ''

TangutComponentDash051

\u{18832}: ''

TangutComponentDash052

\u{18833}: ''

TangutComponentDash053

\u{18834}: ''

TangutComponentDash054

\u{18835}: ''

TangutComponentDash055

\u{18836}: ''

TangutComponentDash056

\u{18837}: ''

TangutComponentDash057

\u{18838}: ''

TangutComponentDash058

\u{18839}: ''

TangutComponentDash059

\u{1883a}: ''

TangutComponentDash060

\u{1883b}: ''

TangutComponentDash061

\u{1883c}: ''

TangutComponentDash062

\u{1883d}: ''

TangutComponentDash063

\u{1883e}: ''

TangutComponentDash064

\u{1883f}: ''

TangutComponentDash065

\u{18840}: ''

TangutComponentDash066

\u{18841}: ''

TangutComponentDash067

\u{18842}: ''

TangutComponentDash068

\u{18843}: ''

TangutComponentDash069

\u{18844}: ''

TangutComponentDash070

\u{18845}: ''

TangutComponentDash071

\u{18846}: ''

TangutComponentDash072

\u{18847}: ''

TangutComponentDash073

\u{18848}: ''

TangutComponentDash074

\u{18849}: ''

TangutComponentDash075

\u{1884a}: ''

TangutComponentDash076

\u{1884b}: ''

TangutComponentDash077

\u{1884c}: ''

TangutComponentDash078

\u{1884d}: ''

TangutComponentDash079

\u{1884e}: ''

TangutComponentDash080

\u{1884f}: ''

TangutComponentDash081

\u{18850}: ''

TangutComponentDash082

\u{18851}: ''

TangutComponentDash083

\u{18852}: ''

TangutComponentDash084

\u{18853}: ''

TangutComponentDash085

\u{18854}: ''

TangutComponentDash086

\u{18855}: ''

TangutComponentDash087

\u{18856}: ''

TangutComponentDash088

\u{18857}: ''

TangutComponentDash089

\u{18858}: ''

TangutComponentDash090

\u{18859}: ''

TangutComponentDash091

\u{1885a}: ''

TangutComponentDash092

\u{1885b}: ''

TangutComponentDash093

\u{1885c}: ''

TangutComponentDash094

\u{1885d}: ''

TangutComponentDash095

\u{1885e}: ''

TangutComponentDash096

\u{1885f}: ''

TangutComponentDash097

\u{18860}: ''

TangutComponentDash098

\u{18861}: ''

TangutComponentDash099

\u{18862}: ''

TangutComponentDash100

\u{18863}: ''

TangutComponentDash101

\u{18864}: ''

TangutComponentDash102

\u{18865}: ''

TangutComponentDash103

\u{18866}: ''

TangutComponentDash104

\u{18867}: ''

TangutComponentDash105

\u{18868}: ''

TangutComponentDash106

\u{18869}: ''

TangutComponentDash107

\u{1886a}: ''

TangutComponentDash108

\u{1886b}: ''

TangutComponentDash109

\u{1886c}: ''

TangutComponentDash110

\u{1886d}: ''

TangutComponentDash111

\u{1886e}: ''

TangutComponentDash112

\u{1886f}: ''

TangutComponentDash113

\u{18870}: ''

TangutComponentDash114

\u{18871}: ''

TangutComponentDash115

\u{18872}: ''

TangutComponentDash116

\u{18873}: ''

TangutComponentDash117

\u{18874}: ''

TangutComponentDash118

\u{18875}: ''

TangutComponentDash119

\u{18876}: ''

TangutComponentDash120

\u{18877}: ''

TangutComponentDash121

\u{18878}: ''

TangutComponentDash122

\u{18879}: ''

TangutComponentDash123

\u{1887a}: ''

TangutComponentDash124

\u{1887b}: ''

TangutComponentDash125

\u{1887c}: ''

TangutComponentDash126

\u{1887d}: ''

TangutComponentDash127

\u{1887e}: ''

TangutComponentDash128

\u{1887f}: ''

TangutComponentDash129

\u{18880}: ''

TangutComponentDash130

\u{18881}: ''

TangutComponentDash131

\u{18882}: ''

TangutComponentDash132

\u{18883}: ''

TangutComponentDash133

\u{18884}: ''

TangutComponentDash134

\u{18885}: ''

TangutComponentDash135

\u{18886}: ''

TangutComponentDash136

\u{18887}: ''

TangutComponentDash137

\u{18888}: ''

TangutComponentDash138

\u{18889}: ''

TangutComponentDash139

\u{1888a}: ''

TangutComponentDash140

\u{1888b}: ''

TangutComponentDash141

\u{1888c}: ''

TangutComponentDash142

\u{1888d}: ''

TangutComponentDash143

\u{1888e}: ''

TangutComponentDash144

\u{1888f}: ''

TangutComponentDash145

\u{18890}: ''

TangutComponentDash146

\u{18891}: ''

TangutComponentDash147

\u{18892}: ''

TangutComponentDash148

\u{18893}: ''

TangutComponentDash149

\u{18894}: ''

TangutComponentDash150

\u{18895}: ''

TangutComponentDash151

\u{18896}: ''

TangutComponentDash152

\u{18897}: ''

TangutComponentDash153

\u{18898}: ''

TangutComponentDash154

\u{18899}: ''

TangutComponentDash155

\u{1889a}: ''

TangutComponentDash156

\u{1889b}: ''

TangutComponentDash157

\u{1889c}: ''

TangutComponentDash158

\u{1889d}: ''

TangutComponentDash159

\u{1889e}: ''

TangutComponentDash160

\u{1889f}: ''

TangutComponentDash161

\u{188a0}: ''

TangutComponentDash162

\u{188a1}: ''

TangutComponentDash163

\u{188a2}: ''

TangutComponentDash164

\u{188a3}: ''

TangutComponentDash165

\u{188a4}: ''

TangutComponentDash166

\u{188a5}: ''

TangutComponentDash167

\u{188a6}: ''

TangutComponentDash168

\u{188a7}: ''

TangutComponentDash169

\u{188a8}: ''

TangutComponentDash170

\u{188a9}: ''

TangutComponentDash171

\u{188aa}: ''

TangutComponentDash172

\u{188ab}: ''

TangutComponentDash173

\u{188ac}: ''

TangutComponentDash174

\u{188ad}: ''

TangutComponentDash175

\u{188ae}: ''

TangutComponentDash176

\u{188af}: ''

TangutComponentDash177

\u{188b0}: ''

TangutComponentDash178

\u{188b1}: ''

TangutComponentDash179

\u{188b2}: ''

TangutComponentDash180

\u{188b3}: ''

TangutComponentDash181

\u{188b4}: ''

TangutComponentDash182

\u{188b5}: ''

TangutComponentDash183

\u{188b6}: ''

TangutComponentDash184

\u{188b7}: ''

TangutComponentDash185

\u{188b8}: ''

TangutComponentDash186

\u{188b9}: ''

TangutComponentDash187

\u{188ba}: ''

TangutComponentDash188

\u{188bb}: ''

TangutComponentDash189

\u{188bc}: ''

TangutComponentDash190

\u{188bd}: ''

TangutComponentDash191

\u{188be}: ''

TangutComponentDash192

\u{188bf}: ''

TangutComponentDash193

\u{188c0}: ''

TangutComponentDash194

\u{188c1}: ''

TangutComponentDash195

\u{188c2}: ''

TangutComponentDash196

\u{188c3}: ''

TangutComponentDash197

\u{188c4}: ''

TangutComponentDash198

\u{188c5}: ''

TangutComponentDash199

\u{188c6}: ''

TangutComponentDash200

\u{188c7}: ''

TangutComponentDash201

\u{188c8}: ''

TangutComponentDash202

\u{188c9}: ''

TangutComponentDash203

\u{188ca}: ''

TangutComponentDash204

\u{188cb}: ''

TangutComponentDash205

\u{188cc}: ''

TangutComponentDash206

\u{188cd}: ''

TangutComponentDash207

\u{188ce}: ''

TangutComponentDash208

\u{188cf}: ''

TangutComponentDash209

\u{188d0}: ''

TangutComponentDash210

\u{188d1}: ''

TangutComponentDash211

\u{188d2}: ''

TangutComponentDash212

\u{188d3}: ''

TangutComponentDash213

\u{188d4}: ''

TangutComponentDash214

\u{188d5}: ''

TangutComponentDash215

\u{188d6}: ''

TangutComponentDash216

\u{188d7}: ''

TangutComponentDash217

\u{188d8}: ''

TangutComponentDash218

\u{188d9}: ''

TangutComponentDash219

\u{188da}: ''

TangutComponentDash220

\u{188db}: ''

TangutComponentDash221

\u{188dc}: ''

TangutComponentDash222

\u{188dd}: ''

TangutComponentDash223

\u{188de}: ''

TangutComponentDash224

\u{188df}: ''

TangutComponentDash225

\u{188e0}: ''

TangutComponentDash226

\u{188e1}: ''

TangutComponentDash227

\u{188e2}: ''

TangutComponentDash228

\u{188e3}: ''

TangutComponentDash229

\u{188e4}: ''

TangutComponentDash230

\u{188e5}: ''

TangutComponentDash231

\u{188e6}: ''

TangutComponentDash232

\u{188e7}: ''

TangutComponentDash233

\u{188e8}: ''

TangutComponentDash234

\u{188e9}: ''

TangutComponentDash235

\u{188ea}: ''

TangutComponentDash236

\u{188eb}: ''

TangutComponentDash237

\u{188ec}: ''

TangutComponentDash238

\u{188ed}: ''

TangutComponentDash239

\u{188ee}: ''

TangutComponentDash240

\u{188ef}: ''

TangutComponentDash241

\u{188f0}: ''

TangutComponentDash242

\u{188f1}: ''

TangutComponentDash243

\u{188f2}: ''

TangutComponentDash244

\u{188f3}: ''

TangutComponentDash245

\u{188f4}: ''

TangutComponentDash246

\u{188f5}: ''

TangutComponentDash247

\u{188f6}: ''

TangutComponentDash248

\u{188f7}: ''

TangutComponentDash249

\u{188f8}: ''

TangutComponentDash250

\u{188f9}: ''

TangutComponentDash251

\u{188fa}: ''

TangutComponentDash252

\u{188fb}: ''

TangutComponentDash253

\u{188fc}: ''

TangutComponentDash254

\u{188fd}: ''

TangutComponentDash255

\u{188fe}: ''

TangutComponentDash256

\u{188ff}: ''

TangutComponentDash257

\u{18900}: ''

TangutComponentDash258

\u{18901}: ''

TangutComponentDash259

\u{18902}: ''

TangutComponentDash260

\u{18903}: ''

TangutComponentDash261

\u{18904}: ''

TangutComponentDash262

\u{18905}: ''

TangutComponentDash263

\u{18906}: ''

TangutComponentDash264

\u{18907}: ''

TangutComponentDash265

\u{18908}: ''

TangutComponentDash266

\u{18909}: ''

TangutComponentDash267

\u{1890a}: ''

TangutComponentDash268

\u{1890b}: ''

TangutComponentDash269

\u{1890c}: ''

TangutComponentDash270

\u{1890d}: ''

TangutComponentDash271

\u{1890e}: ''

TangutComponentDash272

\u{1890f}: ''

TangutComponentDash273

\u{18910}: ''

TangutComponentDash274

\u{18911}: ''

TangutComponentDash275

\u{18912}: ''

TangutComponentDash276

\u{18913}: ''

TangutComponentDash277

\u{18914}: ''

TangutComponentDash278

\u{18915}: ''

TangutComponentDash279

\u{18916}: ''

TangutComponentDash280

\u{18917}: ''

TangutComponentDash281

\u{18918}: ''

TangutComponentDash282

\u{18919}: ''

TangutComponentDash283

\u{1891a}: ''

TangutComponentDash284

\u{1891b}: ''

TangutComponentDash285

\u{1891c}: ''

TangutComponentDash286

\u{1891d}: ''

TangutComponentDash287

\u{1891e}: ''

TangutComponentDash288

\u{1891f}: ''

TangutComponentDash289

\u{18920}: ''

TangutComponentDash290

\u{18921}: ''

TangutComponentDash291

\u{18922}: ''

TangutComponentDash292

\u{18923}: ''

TangutComponentDash293

\u{18924}: ''

TangutComponentDash294

\u{18925}: ''

TangutComponentDash295

\u{18926}: ''

TangutComponentDash296

\u{18927}: ''

TangutComponentDash297

\u{18928}: ''

TangutComponentDash298

\u{18929}: ''

TangutComponentDash299

\u{1892a}: ''

TangutComponentDash300

\u{1892b}: ''

TangutComponentDash301

\u{1892c}: ''

TangutComponentDash302

\u{1892d}: ''

TangutComponentDash303

\u{1892e}: ''

TangutComponentDash304

\u{1892f}: ''

TangutComponentDash305

\u{18930}: ''

TangutComponentDash306

\u{18931}: ''

TangutComponentDash307

\u{18932}: ''

TangutComponentDash308

\u{18933}: ''

TangutComponentDash309

\u{18934}: ''

TangutComponentDash310

\u{18935}: ''

TangutComponentDash311

\u{18936}: ''

TangutComponentDash312

\u{18937}: ''

TangutComponentDash313

\u{18938}: ''

TangutComponentDash314

\u{18939}: ''

TangutComponentDash315

\u{1893a}: ''

TangutComponentDash316

\u{1893b}: ''

TangutComponentDash317

\u{1893c}: ''

TangutComponentDash318

\u{1893d}: ''

TangutComponentDash319

\u{1893e}: ''

TangutComponentDash320

\u{1893f}: ''

TangutComponentDash321

\u{18940}: ''

TangutComponentDash322

\u{18941}: ''

TangutComponentDash323

\u{18942}: ''

TangutComponentDash324

\u{18943}: ''

TangutComponentDash325

\u{18944}: ''

TangutComponentDash326

\u{18945}: ''

TangutComponentDash327

\u{18946}: ''

TangutComponentDash328

\u{18947}: ''

TangutComponentDash329

\u{18948}: ''

TangutComponentDash330

\u{18949}: ''

TangutComponentDash331

\u{1894a}: ''

TangutComponentDash332

\u{1894b}: ''

TangutComponentDash333

\u{1894c}: ''

TangutComponentDash334

\u{1894d}: ''

TangutComponentDash335

\u{1894e}: ''

TangutComponentDash336

\u{1894f}: ''

TangutComponentDash337

\u{18950}: ''

TangutComponentDash338

\u{18951}: ''

TangutComponentDash339

\u{18952}: ''

TangutComponentDash340

\u{18953}: ''

TangutComponentDash341

\u{18954}: ''

TangutComponentDash342

\u{18955}: ''

TangutComponentDash343

\u{18956}: ''

TangutComponentDash344

\u{18957}: ''

TangutComponentDash345

\u{18958}: ''

TangutComponentDash346

\u{18959}: ''

TangutComponentDash347

\u{1895a}: ''

TangutComponentDash348

\u{1895b}: ''

TangutComponentDash349

\u{1895c}: ''

TangutComponentDash350

\u{1895d}: ''

TangutComponentDash351

\u{1895e}: ''

TangutComponentDash352

\u{1895f}: ''

TangutComponentDash353

\u{18960}: ''

TangutComponentDash354

\u{18961}: ''

TangutComponentDash355

\u{18962}: ''

TangutComponentDash356

\u{18963}: ''

TangutComponentDash357

\u{18964}: ''

TangutComponentDash358

\u{18965}: ''

TangutComponentDash359

\u{18966}: ''

TangutComponentDash360

\u{18967}: ''

TangutComponentDash361

\u{18968}: ''

TangutComponentDash362

\u{18969}: ''

TangutComponentDash363

\u{1896a}: ''

TangutComponentDash364

\u{1896b}: ''

TangutComponentDash365

\u{1896c}: ''

TangutComponentDash366

\u{1896d}: ''

TangutComponentDash367

\u{1896e}: ''

TangutComponentDash368

\u{1896f}: ''

TangutComponentDash369

\u{18970}: ''

TangutComponentDash370

\u{18971}: ''

TangutComponentDash371

\u{18972}: ''

TangutComponentDash372

\u{18973}: ''

TangutComponentDash373

\u{18974}: ''

TangutComponentDash374

\u{18975}: ''

TangutComponentDash375

\u{18976}: ''

TangutComponentDash376

\u{18977}: ''

TangutComponentDash377

\u{18978}: ''

TangutComponentDash378

\u{18979}: ''

TangutComponentDash379

\u{1897a}: ''

TangutComponentDash380

\u{1897b}: ''

TangutComponentDash381

\u{1897c}: ''

TangutComponentDash382

\u{1897d}: ''

TangutComponentDash383

\u{1897e}: ''

TangutComponentDash384

\u{1897f}: ''

TangutComponentDash385

\u{18980}: ''

TangutComponentDash386

\u{18981}: ''

TangutComponentDash387

\u{18982}: ''

TangutComponentDash388

\u{18983}: ''

TangutComponentDash389

\u{18984}: ''

TangutComponentDash390

\u{18985}: ''

TangutComponentDash391

\u{18986}: ''

TangutComponentDash392

\u{18987}: ''

TangutComponentDash393

\u{18988}: ''

TangutComponentDash394

\u{18989}: ''

TangutComponentDash395

\u{1898a}: ''

TangutComponentDash396

\u{1898b}: ''

TangutComponentDash397

\u{1898c}: ''

TangutComponentDash398

\u{1898d}: ''

TangutComponentDash399

\u{1898e}: ''

TangutComponentDash400

\u{1898f}: ''

TangutComponentDash401

\u{18990}: ''

TangutComponentDash402

\u{18991}: ''

TangutComponentDash403

\u{18992}: ''

TangutComponentDash404

\u{18993}: ''

TangutComponentDash405

\u{18994}: ''

TangutComponentDash406

\u{18995}: ''

TangutComponentDash407

\u{18996}: ''

TangutComponentDash408

\u{18997}: ''

TangutComponentDash409

\u{18998}: ''

TangutComponentDash410

\u{18999}: ''

TangutComponentDash411

\u{1899a}: ''

TangutComponentDash412

\u{1899b}: ''

TangutComponentDash413

\u{1899c}: ''

TangutComponentDash414

\u{1899d}: ''

TangutComponentDash415

\u{1899e}: ''

TangutComponentDash416

\u{1899f}: ''

TangutComponentDash417

\u{189a0}: ''

TangutComponentDash418

\u{189a1}: ''

TangutComponentDash419

\u{189a2}: ''

TangutComponentDash420

\u{189a3}: ''

TangutComponentDash421

\u{189a4}: ''

TangutComponentDash422

\u{189a5}: ''

TangutComponentDash423

\u{189a6}: ''

TangutComponentDash424

\u{189a7}: ''

TangutComponentDash425

\u{189a8}: ''

TangutComponentDash426

\u{189a9}: ''

TangutComponentDash427

\u{189aa}: ''

TangutComponentDash428

\u{189ab}: ''

TangutComponentDash429

\u{189ac}: ''

TangutComponentDash430

\u{189ad}: ''

TangutComponentDash431

\u{189ae}: ''

TangutComponentDash432

\u{189af}: ''

TangutComponentDash433

\u{189b0}: ''

TangutComponentDash434

\u{189b1}: ''

TangutComponentDash435

\u{189b2}: ''

TangutComponentDash436

\u{189b3}: ''

TangutComponentDash437

\u{189b4}: ''

TangutComponentDash438

\u{189b5}: ''

TangutComponentDash439

\u{189b6}: ''

TangutComponentDash440

\u{189b7}: ''

TangutComponentDash441

\u{189b8}: ''

TangutComponentDash442

\u{189b9}: ''

TangutComponentDash443

\u{189ba}: ''

TangutComponentDash444

\u{189bb}: ''

TangutComponentDash445

\u{189bc}: ''

TangutComponentDash446

\u{189bd}: ''

TangutComponentDash447

\u{189be}: ''

TangutComponentDash448

\u{189bf}: ''

TangutComponentDash449

\u{189c0}: ''

TangutComponentDash450

\u{189c1}: ''

TangutComponentDash451

\u{189c2}: ''

TangutComponentDash452

\u{189c3}: ''

TangutComponentDash453

\u{189c4}: ''

TangutComponentDash454

\u{189c5}: ''

TangutComponentDash455

\u{189c6}: ''

TangutComponentDash456

\u{189c7}: ''

TangutComponentDash457

\u{189c8}: ''

TangutComponentDash458

\u{189c9}: ''

TangutComponentDash459

\u{189ca}: ''

TangutComponentDash460

\u{189cb}: ''

TangutComponentDash461

\u{189cc}: ''

TangutComponentDash462

\u{189cd}: ''

TangutComponentDash463

\u{189ce}: ''

TangutComponentDash464

\u{189cf}: ''

TangutComponentDash465

\u{189d0}: ''

TangutComponentDash466

\u{189d1}: ''

TangutComponentDash467

\u{189d2}: ''

TangutComponentDash468

\u{189d3}: ''

TangutComponentDash469

\u{189d4}: ''

TangutComponentDash470

\u{189d5}: ''

TangutComponentDash471

\u{189d6}: ''

TangutComponentDash472

\u{189d7}: ''

TangutComponentDash473

\u{189d8}: ''

TangutComponentDash474

\u{189d9}: ''

TangutComponentDash475

\u{189da}: ''

TangutComponentDash476

\u{189db}: ''

TangutComponentDash477

\u{189dc}: ''

TangutComponentDash478

\u{189dd}: ''

TangutComponentDash479

\u{189de}: ''

TangutComponentDash480

\u{189df}: ''

TangutComponentDash481

\u{189e0}: ''

TangutComponentDash482

\u{189e1}: ''

TangutComponentDash483

\u{189e2}: ''

TangutComponentDash484

\u{189e3}: ''

TangutComponentDash485

\u{189e4}: ''

TangutComponentDash486

\u{189e5}: ''

TangutComponentDash487

\u{189e6}: ''

TangutComponentDash488

\u{189e7}: ''

TangutComponentDash489

\u{189e8}: ''

TangutComponentDash490

\u{189e9}: ''

TangutComponentDash491

\u{189ea}: ''

TangutComponentDash492

\u{189eb}: ''

TangutComponentDash493

\u{189ec}: ''

TangutComponentDash494

\u{189ed}: ''

TangutComponentDash495

\u{189ee}: ''

TangutComponentDash496

\u{189ef}: ''

TangutComponentDash497

\u{189f0}: ''

TangutComponentDash498

\u{189f1}: ''

TangutComponentDash499

\u{189f2}: ''

TangutComponentDash500

\u{189f3}: ''

TangutComponentDash501

\u{189f4}: ''

TangutComponentDash502

\u{189f5}: ''

TangutComponentDash503

\u{189f6}: ''

TangutComponentDash504

\u{189f7}: ''

TangutComponentDash505

\u{189f8}: ''

TangutComponentDash506

\u{189f9}: ''

TangutComponentDash507

\u{189fa}: ''

TangutComponentDash508

\u{189fb}: ''

TangutComponentDash509

\u{189fc}: ''

TangutComponentDash510

\u{189fd}: ''

TangutComponentDash511

\u{189fe}: ''

TangutComponentDash512

\u{189ff}: ''

TangutComponentDash513

\u{18a00}: ''

TangutComponentDash514

\u{18a01}: ''

TangutComponentDash515

\u{18a02}: ''

TangutComponentDash516

\u{18a03}: ''

TangutComponentDash517

\u{18a04}: ''

TangutComponentDash518

\u{18a05}: ''

TangutComponentDash519

\u{18a06}: ''

TangutComponentDash520

\u{18a07}: ''

TangutComponentDash521

\u{18a08}: ''

TangutComponentDash522

\u{18a09}: ''

TangutComponentDash523

\u{18a0a}: ''

TangutComponentDash524

\u{18a0b}: ''

TangutComponentDash525

\u{18a0c}: ''

TangutComponentDash526

\u{18a0d}: ''

TangutComponentDash527

\u{18a0e}: ''

TangutComponentDash528

\u{18a0f}: ''

TangutComponentDash529

\u{18a10}: ''

TangutComponentDash530

\u{18a11}: ''

TangutComponentDash531

\u{18a12}: ''

TangutComponentDash532

\u{18a13}: ''

TangutComponentDash533

\u{18a14}: ''

TangutComponentDash534

\u{18a15}: ''

TangutComponentDash535

\u{18a16}: ''

TangutComponentDash536

\u{18a17}: ''

TangutComponentDash537

\u{18a18}: ''

TangutComponentDash538

\u{18a19}: ''

TangutComponentDash539

\u{18a1a}: ''

TangutComponentDash540

\u{18a1b}: ''

TangutComponentDash541

\u{18a1c}: ''

TangutComponentDash542

\u{18a1d}: ''

TangutComponentDash543

\u{18a1e}: ''

TangutComponentDash544

\u{18a1f}: ''

TangutComponentDash545

\u{18a20}: ''

TangutComponentDash546

\u{18a21}: ''

TangutComponentDash547

\u{18a22}: ''

TangutComponentDash548

\u{18a23}: ''

TangutComponentDash549

\u{18a24}: ''

TangutComponentDash550

\u{18a25}: ''

TangutComponentDash551

\u{18a26}: ''

TangutComponentDash552

\u{18a27}: ''

TangutComponentDash553

\u{18a28}: ''

TangutComponentDash554

\u{18a29}: ''

TangutComponentDash555

\u{18a2a}: ''

TangutComponentDash556

\u{18a2b}: ''

TangutComponentDash557

\u{18a2c}: ''

TangutComponentDash558

\u{18a2d}: ''

TangutComponentDash559

\u{18a2e}: ''

TangutComponentDash560

\u{18a2f}: ''

TangutComponentDash561

\u{18a30}: ''

TangutComponentDash562

\u{18a31}: ''

TangutComponentDash563

\u{18a32}: ''

TangutComponentDash564

\u{18a33}: ''

TangutComponentDash565

\u{18a34}: ''

TangutComponentDash566

\u{18a35}: ''

TangutComponentDash567

\u{18a36}: ''

TangutComponentDash568

\u{18a37}: ''

TangutComponentDash569

\u{18a38}: ''

TangutComponentDash570

\u{18a39}: ''

TangutComponentDash571

\u{18a3a}: ''

TangutComponentDash572

\u{18a3b}: ''

TangutComponentDash573

\u{18a3c}: ''

TangutComponentDash574

\u{18a3d}: ''

TangutComponentDash575

\u{18a3e}: ''

TangutComponentDash576

\u{18a3f}: ''

TangutComponentDash577

\u{18a40}: ''

TangutComponentDash578

\u{18a41}: ''

TangutComponentDash579

\u{18a42}: ''

TangutComponentDash580

\u{18a43}: ''

TangutComponentDash581

\u{18a44}: ''

TangutComponentDash582

\u{18a45}: ''

TangutComponentDash583

\u{18a46}: ''

TangutComponentDash584

\u{18a47}: ''

TangutComponentDash585

\u{18a48}: ''

TangutComponentDash586

\u{18a49}: ''

TangutComponentDash587

\u{18a4a}: ''

TangutComponentDash588

\u{18a4b}: ''

TangutComponentDash589

\u{18a4c}: ''

TangutComponentDash590

\u{18a4d}: ''

TangutComponentDash591

\u{18a4e}: ''

TangutComponentDash592

\u{18a4f}: ''

TangutComponentDash593

\u{18a50}: ''

TangutComponentDash594

\u{18a51}: ''

TangutComponentDash595

\u{18a52}: ''

TangutComponentDash596

\u{18a53}: ''

TangutComponentDash597

\u{18a54}: ''

TangutComponentDash598

\u{18a55}: ''

TangutComponentDash599

\u{18a56}: ''

TangutComponentDash600

\u{18a57}: ''

TangutComponentDash601

\u{18a58}: ''

TangutComponentDash602

\u{18a59}: ''

TangutComponentDash603

\u{18a5a}: ''

TangutComponentDash604

\u{18a5b}: ''

TangutComponentDash605

\u{18a5c}: ''

TangutComponentDash606

\u{18a5d}: ''

TangutComponentDash607

\u{18a5e}: ''

TangutComponentDash608

\u{18a5f}: ''

TangutComponentDash609

\u{18a60}: ''

TangutComponentDash610

\u{18a61}: ''

TangutComponentDash611

\u{18a62}: ''

TangutComponentDash612

\u{18a63}: ''

TangutComponentDash613

\u{18a64}: ''

TangutComponentDash614

\u{18a65}: ''

TangutComponentDash615

\u{18a66}: ''

TangutComponentDash616

\u{18a67}: ''

TangutComponentDash617

\u{18a68}: ''

TangutComponentDash618

\u{18a69}: ''

TangutComponentDash619

\u{18a6a}: ''

TangutComponentDash620

\u{18a6b}: ''

TangutComponentDash621

\u{18a6c}: ''

TangutComponentDash622

\u{18a6d}: ''

TangutComponentDash623

\u{18a6e}: ''

TangutComponentDash624

\u{18a6f}: ''

TangutComponentDash625

\u{18a70}: ''

TangutComponentDash626

\u{18a71}: ''

TangutComponentDash627

\u{18a72}: ''

TangutComponentDash628

\u{18a73}: ''

TangutComponentDash629

\u{18a74}: ''

TangutComponentDash630

\u{18a75}: ''

TangutComponentDash631

\u{18a76}: ''

TangutComponentDash632

\u{18a77}: ''

TangutComponentDash633

\u{18a78}: ''

TangutComponentDash634

\u{18a79}: ''

TangutComponentDash635

\u{18a7a}: ''

TangutComponentDash636

\u{18a7b}: ''

TangutComponentDash637

\u{18a7c}: ''

TangutComponentDash638

\u{18a7d}: ''

TangutComponentDash639

\u{18a7e}: ''

TangutComponentDash640

\u{18a7f}: ''

TangutComponentDash641

\u{18a80}: ''

TangutComponentDash642

\u{18a81}: ''

TangutComponentDash643

\u{18a82}: ''

TangutComponentDash644

\u{18a83}: ''

TangutComponentDash645

\u{18a84}: ''

TangutComponentDash646

\u{18a85}: ''

TangutComponentDash647

\u{18a86}: ''

TangutComponentDash648

\u{18a87}: ''

TangutComponentDash649

\u{18a88}: ''

TangutComponentDash650

\u{18a89}: ''

TangutComponentDash651

\u{18a8a}: ''

TangutComponentDash652

\u{18a8b}: ''

TangutComponentDash653

\u{18a8c}: ''

TangutComponentDash654

\u{18a8d}: ''

TangutComponentDash655

\u{18a8e}: ''

TangutComponentDash656

\u{18a8f}: ''

TangutComponentDash657

\u{18a90}: ''

TangutComponentDash658

\u{18a91}: ''

TangutComponentDash659

\u{18a92}: ''

TangutComponentDash660

\u{18a93}: ''

TangutComponentDash661

\u{18a94}: ''

TangutComponentDash662

\u{18a95}: ''

TangutComponentDash663

\u{18a96}: ''

TangutComponentDash664

\u{18a97}: ''

TangutComponentDash665

\u{18a98}: ''

TangutComponentDash666

\u{18a99}: ''

TangutComponentDash667

\u{18a9a}: ''

TangutComponentDash668

\u{18a9b}: ''

TangutComponentDash669

\u{18a9c}: ''

TangutComponentDash670

\u{18a9d}: ''

TangutComponentDash671

\u{18a9e}: ''

TangutComponentDash672

\u{18a9f}: ''

TangutComponentDash673

\u{18aa0}: ''

TangutComponentDash674

\u{18aa1}: ''

TangutComponentDash675

\u{18aa2}: ''

TangutComponentDash676

\u{18aa3}: ''

TangutComponentDash677

\u{18aa4}: ''

TangutComponentDash678

\u{18aa5}: ''

TangutComponentDash679

\u{18aa6}: ''

TangutComponentDash680

\u{18aa7}: ''

TangutComponentDash681

\u{18aa8}: ''

TangutComponentDash682

\u{18aa9}: ''

TangutComponentDash683

\u{18aaa}: ''

TangutComponentDash684

\u{18aab}: ''

TangutComponentDash685

\u{18aac}: ''

TangutComponentDash686

\u{18aad}: ''

TangutComponentDash687

\u{18aae}: ''

TangutComponentDash688

\u{18aaf}: ''

TangutComponentDash689

\u{18ab0}: ''

TangutComponentDash690

\u{18ab1}: ''

TangutComponentDash691

\u{18ab2}: ''

TangutComponentDash692

\u{18ab3}: ''

TangutComponentDash693

\u{18ab4}: ''

TangutComponentDash694

\u{18ab5}: ''

TangutComponentDash695

\u{18ab6}: ''

TangutComponentDash696

\u{18ab7}: ''

TangutComponentDash697

\u{18ab8}: ''

TangutComponentDash698

\u{18ab9}: ''

TangutComponentDash699

\u{18aba}: ''

TangutComponentDash700

\u{18abb}: ''

TangutComponentDash701

\u{18abc}: ''

TangutComponentDash702

\u{18abd}: ''

TangutComponentDash703

\u{18abe}: ''

TangutComponentDash704

\u{18abf}: ''

TangutComponentDash705

\u{18ac0}: ''

TangutComponentDash706

\u{18ac1}: ''

TangutComponentDash707

\u{18ac2}: ''

TangutComponentDash708

\u{18ac3}: ''

TangutComponentDash709

\u{18ac4}: ''

TangutComponentDash710

\u{18ac5}: ''

TangutComponentDash711

\u{18ac6}: ''

TangutComponentDash712

\u{18ac7}: ''

TangutComponentDash713

\u{18ac8}: ''

TangutComponentDash714

\u{18ac9}: ''

TangutComponentDash715

\u{18aca}: ''

TangutComponentDash716

\u{18acb}: ''

TangutComponentDash717

\u{18acc}: ''

TangutComponentDash718

\u{18acd}: ''

TangutComponentDash719

\u{18ace}: ''

TangutComponentDash720

\u{18acf}: ''

TangutComponentDash721

\u{18ad0}: ''

TangutComponentDash722

\u{18ad1}: ''

TangutComponentDash723

\u{18ad2}: ''

TangutComponentDash724

\u{18ad3}: ''

TangutComponentDash725

\u{18ad4}: ''

TangutComponentDash726

\u{18ad5}: ''

TangutComponentDash727

\u{18ad6}: ''

TangutComponentDash728

\u{18ad7}: ''

TangutComponentDash729

\u{18ad8}: ''

TangutComponentDash730

\u{18ad9}: ''

TangutComponentDash731

\u{18ada}: ''

TangutComponentDash732

\u{18adb}: ''

TangutComponentDash733

\u{18adc}: ''

TangutComponentDash734

\u{18add}: ''

TangutComponentDash735

\u{18ade}: ''

TangutComponentDash736

\u{18adf}: ''

TangutComponentDash737

\u{18ae0}: ''

TangutComponentDash738

\u{18ae1}: ''

TangutComponentDash739

\u{18ae2}: ''

TangutComponentDash740

\u{18ae3}: ''

TangutComponentDash741

\u{18ae4}: ''

TangutComponentDash742

\u{18ae5}: ''

TangutComponentDash743

\u{18ae6}: ''

TangutComponentDash744

\u{18ae7}: ''

TangutComponentDash745

\u{18ae8}: ''

TangutComponentDash746

\u{18ae9}: ''

TangutComponentDash747

\u{18aea}: ''

TangutComponentDash748

\u{18aeb}: ''

TangutComponentDash749

\u{18aec}: ''

TangutComponentDash750

\u{18aed}: ''

TangutComponentDash751

\u{18aee}: ''

TangutComponentDash752

\u{18aef}: ''

TangutComponentDash753

\u{18af0}: ''

TangutComponentDash754

\u{18af1}: ''

TangutComponentDash755

\u{18af2}: ''

Methods

impl TangutComponents

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case