Enum unicode_types::transport_and_map_symbols::TransportandMapSymbols

pub enum TransportandMapSymbols {
    Rocket,
    Helicopter,
    SteamLocomotive,
    RailwayCar,
    HighDashSpeedTrain,
    HighDashSpeedTrainWithBulletNose,
    Train,
    Metro,
    LightRail,
    Station,
    Tram,
    TramCar,
    Bus,
    OncomingBus,
    Trolleybus,
    BusStop,
    Minibus,
    Ambulance,
    FireEngine,
    PoliceCar,
    OncomingPoliceCar,
    Taxi,
    OncomingTaxi,
    Automobile,
    OncomingAutomobile,
    RecreationalVehicle,
    DeliveryTruck,
    ArticulatedLorry,
    Tractor,
    Monorail,
    MountainRailway,
    SuspensionRailway,
    MountainCableway,
    AerialTramway,
    Ship,
    Rowboat,
    Speedboat,
    HorizontalTrafficLight,
    VerticalTrafficLight,
    ConstructionSign,
    PoliceCarsRevolvingLight,
    TriangularFlagOnPost,
    Door,
    NoEntrySign,
    SmokingSymbol,
    NoSmokingSymbol,
    PutLitterInItsPlaceSymbol,
    DoNotLitterSymbol,
    PotableWaterSymbol,
    NonDashPotableWaterSymbol,
    Bicycle,
    NoBicycles,
    Bicyclist,
    MountainBicyclist,
    Pedestrian,
    NoPedestrians,
    ChildrenCrossing,
    MensSymbol,
    WomensSymbol,
    Restroom,
    BabySymbol,
    Toilet,
    WaterCloset,
    Shower,
    Bath,
    Bathtub,
    PassportControl,
    Customs,
    BaggageClaim,
    LeftLuggage,
    TriangleWithRoundedCorners,
    ProhibitedSign,
    CircledInformationSource,
    BoysSymbol,
    GirlsSymbol,
    CouchAndLamp,
    SleepingAccommodation,
    ShoppingBags,
    BellhopBell,
    Bed,
    PlaceOfWorship,
    OctagonalSign,
    ShoppingTrolley,
    Stupa,
    Pagoda,
    HinduTemple,
    HammerAndWrench,
    Shield,
    OilDrum,
    Motorway,
    RailwayTrack,
    MotorBoat,
    UpDashPointingMilitaryAirplane,
    UpDashPointingAirplane,
    UpDashPointingSmallAirplane,
    SmallAirplane,
    NortheastDashPointingAirplane,
    AirplaneDeparture,
    AirplaneArriving,
    Satellite,
    OncomingFireEngine,
    DieselLocomotive,
    PassengerShip,
    Scooter,
    MotorScooter,
    Canoe,
    Sled,
    FlyingSaucer,
    Skateboard,
    AutoRickshaw,
}

An enum to represent all characters in the TransportandMapSymbols block.

Variants
Rocket

\u{1f680}: 'ğŸš€'

Helicopter

\u{1f681}: 'ğŸš'

SteamLocomotive

\u{1f682}: 'ğŸš‚'

RailwayCar

\u{1f683}: 'ğŸšƒ'

HighDashSpeedTrain

\u{1f684}: 'ğŸš„'

HighDashSpeedTrainWithBulletNose

\u{1f685}: 'ğŸš…'

Train

\u{1f686}: 'ğŸš†'

Metro

\u{1f687}: 'ğŸš‡'

LightRail

\u{1f688}: 'ğŸšˆ'

Station

\u{1f689}: 'ğŸš‰'

Tram

\u{1f68a}: 'ğŸšŠ'

TramCar

\u{1f68b}: 'ğŸš‹'

Bus

\u{1f68c}: 'ğŸšŒ'

OncomingBus

\u{1f68d}: 'ğŸš'

Trolleybus

\u{1f68e}: 'ğŸš'

BusStop

\u{1f68f}: 'ğŸš'

Minibus

\u{1f690}: 'ğŸš'

Ambulance

\u{1f691}: 'ğŸš‘'

FireEngine

\u{1f692}: 'ğŸš’'

PoliceCar

\u{1f693}: 'ğŸš“'

OncomingPoliceCar

\u{1f694}: 'ğŸš”'

Taxi

\u{1f695}: 'ğŸš•'

OncomingTaxi

\u{1f696}: 'ğŸš–'

Automobile

\u{1f697}: 'ğŸš—'

OncomingAutomobile

\u{1f698}: 'ğŸš˜'

RecreationalVehicle

\u{1f699}: 'ğŸš™'

DeliveryTruck

\u{1f69a}: 'ğŸšš'

ArticulatedLorry

\u{1f69b}: 'ğŸš›'

Tractor

\u{1f69c}: 'ğŸšœ'

Monorail

\u{1f69d}: 'ğŸš'

MountainRailway

\u{1f69e}: 'ğŸš'

SuspensionRailway

\u{1f69f}: 'ğŸšŸ'

MountainCableway

\u{1f6a0}: 'ğŸš '

AerialTramway

\u{1f6a1}: 'ğŸš¡'

Ship

\u{1f6a2}: 'ğŸš¢'

Rowboat

\u{1f6a3}: 'ğŸš£'

Speedboat

\u{1f6a4}: 'ğŸš¤'

HorizontalTrafficLight

\u{1f6a5}: 'ğŸš¥'

VerticalTrafficLight

\u{1f6a6}: 'ğŸš¦'

ConstructionSign

\u{1f6a7}: 'ğŸš'

PoliceCarsRevolvingLight

\u{1f6a8}: 'ğŸš¨'

TriangularFlagOnPost

\u{1f6a9}: 'ğŸš©'

Door

\u{1f6aa}: 'ğŸšª'

NoEntrySign

\u{1f6ab}: 'ğŸš«'

SmokingSymbol

\u{1f6ac}: 'ğŸš¬'

NoSmokingSymbol

\u{1f6ad}: 'ğŸš­'

PutLitterInItsPlaceSymbol

\u{1f6ae}: 'ğŸš®'

DoNotLitterSymbol

\u{1f6af}: 'ğŸš¯'

PotableWaterSymbol

\u{1f6b0}: 'ğŸš°'

NonDashPotableWaterSymbol

\u{1f6b1}: 'ğŸš±'

Bicycle

\u{1f6b2}: 'ğŸš²'

NoBicycles

\u{1f6b3}: 'ğŸš³'

Bicyclist

\u{1f6b4}: 'ğŸš´'

MountainBicyclist

\u{1f6b5}: 'ğŸšµ'

Pedestrian

\u{1f6b6}: 'ğŸš¶'

NoPedestrians

\u{1f6b7}: 'ğŸš·'

ChildrenCrossing

\u{1f6b8}: 'ğŸš¸'

MensSymbol

\u{1f6b9}: 'ğŸš¹'

WomensSymbol

\u{1f6ba}: 'ğŸšº'

Restroom

\u{1f6bb}: 'ğŸš»'

BabySymbol

\u{1f6bc}: 'ğŸš¼'

Toilet

\u{1f6bd}: 'ğŸš½'

WaterCloset

\u{1f6be}: 'ğŸš¾'

Shower

\u{1f6bf}: 'ğŸš¿'

Bath

\u{1f6c0}: 'ğŸ›€'

Bathtub

\u{1f6c1}: 'ğŸ›'

PassportControl

\u{1f6c2}: 'ğŸ›‚'

Customs

\u{1f6c3}: 'ğŸ›ƒ'

BaggageClaim

\u{1f6c4}: 'ğŸ›„'

LeftLuggage

\u{1f6c5}: 'ğŸ›…'

TriangleWithRoundedCorners

\u{1f6c6}: 'ğŸ›†'

ProhibitedSign

\u{1f6c7}: 'ğŸ›‡'

CircledInformationSource

\u{1f6c8}: 'ğŸ›ˆ'

BoysSymbol

\u{1f6c9}: 'ğŸ›‰'

GirlsSymbol

\u{1f6ca}: 'ğŸ›Š'

CouchAndLamp

\u{1f6cb}: 'ğŸ›‹'

SleepingAccommodation

\u{1f6cc}: 'ğŸ›Œ'

ShoppingBags

\u{1f6cd}: 'ğŸ›'

BellhopBell

\u{1f6ce}: 'ğŸ›'

Bed

\u{1f6cf}: 'ğŸ›'

PlaceOfWorship

\u{1f6d0}: 'ğŸ›'

OctagonalSign

\u{1f6d1}: 'ğŸ›‘'

ShoppingTrolley

\u{1f6d2}: 'ğŸ›’'

Stupa

\u{1f6d3}: 'ğŸ›“'

Pagoda

\u{1f6d4}: 'ğŸ›”'

HinduTemple

\u{1f6d5}: 'ğŸ›•'

HammerAndWrench

\u{1f6e0}: 'ğŸ› '

Shield

\u{1f6e1}: 'ğŸ›¡'

OilDrum

\u{1f6e2}: 'ğŸ›¢'

Motorway

\u{1f6e3}: 'ğŸ›£'

RailwayTrack

\u{1f6e4}: 'ğŸ›¤'

MotorBoat

\u{1f6e5}: 'ğŸ›¥'

UpDashPointingMilitaryAirplane

\u{1f6e6}: 'ğŸ›¦'

UpDashPointingAirplane

\u{1f6e7}: 'ğŸ›'

UpDashPointingSmallAirplane

\u{1f6e8}: 'ğŸ›¨'

SmallAirplane

\u{1f6e9}: 'ğŸ›©'

NortheastDashPointingAirplane

\u{1f6ea}: 'ğŸ›ª'

AirplaneDeparture

\u{1f6eb}: 'ğŸ›«'

AirplaneArriving

\u{1f6ec}: 'ğŸ›¬'

Satellite

\u{1f6f0}: 'ğŸ›°'

OncomingFireEngine

\u{1f6f1}: 'ğŸ›±'

DieselLocomotive

\u{1f6f2}: 'ğŸ›²'

PassengerShip

\u{1f6f3}: 'ğŸ›³'

Scooter

\u{1f6f4}: 'ğŸ›´'

MotorScooter

\u{1f6f5}: 'ğŸ›µ'

Canoe

\u{1f6f6}: 'ğŸ›¶'

Sled

\u{1f6f7}: 'ğŸ›·'

FlyingSaucer

\u{1f6f8}: 'ğŸ›¸'

Skateboard

\u{1f6f9}: 'ğŸ›¹'

AutoRickshaw

\u{1f6fa}: 'ğŸ›º'

Methods

impl TransportandMapSymbols

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case