Enum unicode_types::unified_canadian_aboriginal_syllabics::UnifiedCanadianAboriginalSyllabics

pub enum UnifiedCanadianAboriginalSyllabics {
    CanadianSyllabicsHyphen,
    CanadianSyllabicsE,
    CanadianSyllabicsAai,
    CanadianSyllabicsI,
    CanadianSyllabicsIi,
    CanadianSyllabicsO,
    CanadianSyllabicsOo,
    CanadianSyllabicsYDashCreeOo,
    CanadianSyllabicsCarrierEe,
    CanadianSyllabicsCarrierI,
    CanadianSyllabicsA,
    CanadianSyllabicsAa,
    CanadianSyllabicsWe,
    CanadianSyllabicsWestDashCreeWe,
    CanadianSyllabicsWi,
    CanadianSyllabicsWestDashCreeWi,
    CanadianSyllabicsWii,
    CanadianSyllabicsWestDashCreeWii,
    CanadianSyllabicsWo,
    CanadianSyllabicsWestDashCreeWo,
    CanadianSyllabicsWoo,
    CanadianSyllabicsWestDashCreeWoo,
    CanadianSyllabicsNaskapiWoo,
    CanadianSyllabicsWa,
    CanadianSyllabicsWestDashCreeWa,
    CanadianSyllabicsWaa,
    CanadianSyllabicsWestDashCreeWaa,
    CanadianSyllabicsNaskapiWaa,
    CanadianSyllabicsAi,
    CanadianSyllabicsYDashCreeW,
    CanadianSyllabicsGlottalStop,
    CanadianSyllabicsFinalAcute,
    CanadianSyllabicsFinalGrave,
    CanadianSyllabicsFinalBottomHalfRing,
    CanadianSyllabicsFinalTopHalfRing,
    CanadianSyllabicsFinalRightHalfRing,
    CanadianSyllabicsFinalRing,
    CanadianSyllabicsFinalDoubleAcute,
    CanadianSyllabicsFinalDoubleShortVerticalStrokes,
    CanadianSyllabicsFinalMiddleDot,
    CanadianSyllabicsFinalShortHorizontalStroke,
    CanadianSyllabicsFinalPlus,
    CanadianSyllabicsFinalDownTack,
    CanadianSyllabicsEn,
    CanadianSyllabicsIn,
    CanadianSyllabicsOn,
    CanadianSyllabicsAn,
    CanadianSyllabicsPe,
    CanadianSyllabicsPaai,
    CanadianSyllabicsPi,
    CanadianSyllabicsPii,
    CanadianSyllabicsPo,
    CanadianSyllabicsPoo,
    CanadianSyllabicsYDashCreePoo,
    CanadianSyllabicsCarrierHee,
    CanadianSyllabicsCarrierHi,
    CanadianSyllabicsPa,
    CanadianSyllabicsPaa,
    CanadianSyllabicsPwe,
    CanadianSyllabicsWestDashCreePwe,
    CanadianSyllabicsPwi,
    CanadianSyllabicsWestDashCreePwi,
    CanadianSyllabicsPwii,
    CanadianSyllabicsWestDashCreePwii,
    CanadianSyllabicsPwo,
    CanadianSyllabicsWestDashCreePwo,
    CanadianSyllabicsPwoo,
    CanadianSyllabicsWestDashCreePwoo,
    CanadianSyllabicsPwa,
    CanadianSyllabicsWestDashCreePwa,
    CanadianSyllabicsPwaa,
    CanadianSyllabicsWestDashCreePwaa,
    CanadianSyllabicsYDashCreePwaa,
    CanadianSyllabicsP,
    CanadianSyllabicsWestDashCreeP,
    CanadianSyllabicsCarrierH,
    CanadianSyllabicsTe,
    CanadianSyllabicsTaai,
    CanadianSyllabicsTi,
    CanadianSyllabicsTii,
    CanadianSyllabicsTo,
    CanadianSyllabicsToo,
    CanadianSyllabicsYDashCreeToo,
    CanadianSyllabicsCarrierDee,
    CanadianSyllabicsCarrierDi,
    CanadianSyllabicsTa,
    CanadianSyllabicsTaa,
    CanadianSyllabicsTwe,
    CanadianSyllabicsWestDashCreeTwe,
    CanadianSyllabicsTwi,
    CanadianSyllabicsWestDashCreeTwi,
    CanadianSyllabicsTwii,
    CanadianSyllabicsWestDashCreeTwii,
    CanadianSyllabicsTwo,
    CanadianSyllabicsWestDashCreeTwo,
    CanadianSyllabicsTwoo,
    CanadianSyllabicsWestDashCreeTwoo,
    CanadianSyllabicsTwa,
    CanadianSyllabicsWestDashCreeTwa,
    CanadianSyllabicsTwaa,
    CanadianSyllabicsWestDashCreeTwaa,
    CanadianSyllabicsNaskapiTwaa,
    CanadianSyllabicsT,
    CanadianSyllabicsTte,
    CanadianSyllabicsTti,
    CanadianSyllabicsTto,
    CanadianSyllabicsTta,
    CanadianSyllabicsKe,
    CanadianSyllabicsKaai,
    CanadianSyllabicsKi,
    CanadianSyllabicsKii,
    CanadianSyllabicsKo,
    CanadianSyllabicsKoo,
    CanadianSyllabicsYDashCreeKoo,
    CanadianSyllabicsKa,
    CanadianSyllabicsKaa,
    CanadianSyllabicsKwe,
    CanadianSyllabicsWestDashCreeKwe,
    CanadianSyllabicsKwi,
    CanadianSyllabicsWestDashCreeKwi,
    CanadianSyllabicsKwii,
    CanadianSyllabicsWestDashCreeKwii,
    CanadianSyllabicsKwo,
    CanadianSyllabicsWestDashCreeKwo,
    CanadianSyllabicsKwoo,
    CanadianSyllabicsWestDashCreeKwoo,
    CanadianSyllabicsKwa,
    CanadianSyllabicsWestDashCreeKwa,
    CanadianSyllabicsKwaa,
    CanadianSyllabicsWestDashCreeKwaa,
    CanadianSyllabicsNaskapiKwaa,
    CanadianSyllabicsK,
    CanadianSyllabicsKw,
    CanadianSyllabicsSouthDashSlaveyKeh,
    CanadianSyllabicsSouthDashSlaveyKih,
    CanadianSyllabicsSouthDashSlaveyKoh,
    CanadianSyllabicsSouthDashSlaveyKah,
    CanadianSyllabicsCe,
    CanadianSyllabicsCaai,
    CanadianSyllabicsCi,
    CanadianSyllabicsCii,
    CanadianSyllabicsCo,
    CanadianSyllabicsCoo,
    CanadianSyllabicsYDashCreeCoo,
    CanadianSyllabicsCa,
    CanadianSyllabicsCaa,
    CanadianSyllabicsCwe,
    CanadianSyllabicsWestDashCreeCwe,
    CanadianSyllabicsCwi,
    CanadianSyllabicsWestDashCreeCwi,
    CanadianSyllabicsCwii,
    CanadianSyllabicsWestDashCreeCwii,
    CanadianSyllabicsCwo,
    CanadianSyllabicsWestDashCreeCwo,
    CanadianSyllabicsCwoo,
    CanadianSyllabicsWestDashCreeCwoo,
    CanadianSyllabicsCwa,
    CanadianSyllabicsWestDashCreeCwa,
    CanadianSyllabicsCwaa,
    CanadianSyllabicsWestDashCreeCwaa,
    CanadianSyllabicsNaskapiCwaa,
    CanadianSyllabicsC,
    CanadianSyllabicsSayisiTh,
    CanadianSyllabicsMe,
    CanadianSyllabicsMaai,
    CanadianSyllabicsMi,
    CanadianSyllabicsMii,
    CanadianSyllabicsMo,
    CanadianSyllabicsMoo,
    CanadianSyllabicsYDashCreeMoo,
    CanadianSyllabicsMa,
    CanadianSyllabicsMaa,
    CanadianSyllabicsMwe,
    CanadianSyllabicsWestDashCreeMwe,
    CanadianSyllabicsMwi,
    CanadianSyllabicsWestDashCreeMwi,
    CanadianSyllabicsMwii,
    CanadianSyllabicsWestDashCreeMwii,
    CanadianSyllabicsMwo,
    CanadianSyllabicsWestDashCreeMwo,
    CanadianSyllabicsMwoo,
    CanadianSyllabicsWestDashCreeMwoo,
    CanadianSyllabicsMwa,
    CanadianSyllabicsWestDashCreeMwa,
    CanadianSyllabicsMwaa,
    CanadianSyllabicsWestDashCreeMwaa,
    CanadianSyllabicsNaskapiMwaa,
    CanadianSyllabicsM,
    CanadianSyllabicsWestDashCreeM,
    CanadianSyllabicsMh,
    CanadianSyllabicsAthapascanM,
    CanadianSyllabicsSayisiM,
    CanadianSyllabicsNe,
    CanadianSyllabicsNaai,
    CanadianSyllabicsNi,
    CanadianSyllabicsNii,
    CanadianSyllabicsNo,
    CanadianSyllabicsNoo,
    CanadianSyllabicsYDashCreeNoo,
    CanadianSyllabicsNa,
    CanadianSyllabicsNaa,
    CanadianSyllabicsNwe,
    CanadianSyllabicsWestDashCreeNwe,
    CanadianSyllabicsNwa,
    CanadianSyllabicsWestDashCreeNwa,
    CanadianSyllabicsNwaa,
    CanadianSyllabicsWestDashCreeNwaa,
    CanadianSyllabicsNaskapiNwaa,
    CanadianSyllabicsN,
    CanadianSyllabicsCarrierNg,
    CanadianSyllabicsNh,
    CanadianSyllabicsLe,
    CanadianSyllabicsLaai,
    CanadianSyllabicsLi,
    CanadianSyllabicsLii,
    CanadianSyllabicsLo,
    CanadianSyllabicsLoo,
    CanadianSyllabicsYDashCreeLoo,
    CanadianSyllabicsLa,
    CanadianSyllabicsLaa,
    CanadianSyllabicsLwe,
    CanadianSyllabicsWestDashCreeLwe,
    CanadianSyllabicsLwi,
    CanadianSyllabicsWestDashCreeLwi,
    CanadianSyllabicsLwii,
    CanadianSyllabicsWestDashCreeLwii,
    CanadianSyllabicsLwo,
    CanadianSyllabicsWestDashCreeLwo,
    CanadianSyllabicsLwoo,
    CanadianSyllabicsWestDashCreeLwoo,
    CanadianSyllabicsLwa,
    CanadianSyllabicsWestDashCreeLwa,
    CanadianSyllabicsLwaa,
    CanadianSyllabicsWestDashCreeLwaa,
    CanadianSyllabicsL,
    CanadianSyllabicsWestDashCreeL,
    CanadianSyllabicsMedialL,
    CanadianSyllabicsSe,
    CanadianSyllabicsSaai,
    CanadianSyllabicsSi,
    CanadianSyllabicsSii,
    CanadianSyllabicsSo,
    CanadianSyllabicsSoo,
    CanadianSyllabicsYDashCreeSoo,
    CanadianSyllabicsSa,
    CanadianSyllabicsSaa,
    CanadianSyllabicsSwe,
    CanadianSyllabicsWestDashCreeSwe,
    CanadianSyllabicsSwi,
    CanadianSyllabicsWestDashCreeSwi,
    CanadianSyllabicsSwii,
    CanadianSyllabicsWestDashCreeSwii,
    CanadianSyllabicsSwo,
    CanadianSyllabicsWestDashCreeSwo,
    CanadianSyllabicsSwoo,
    CanadianSyllabicsWestDashCreeSwoo,
    CanadianSyllabicsSwa,
    CanadianSyllabicsWestDashCreeSwa,
    CanadianSyllabicsSwaa,
    CanadianSyllabicsWestDashCreeSwaa,
    CanadianSyllabicsNaskapiSwaa,
    CanadianSyllabicsS,
    CanadianSyllabicsAthapascanS,
    CanadianSyllabicsSw,
    CanadianSyllabicsBlackfootS,
    CanadianSyllabicsMooseDashCreeSk,
    CanadianSyllabicsNaskapiSkw,
    CanadianSyllabicsNaskapiSDashW,
    CanadianSyllabicsNaskapiSpwa,
    CanadianSyllabicsNaskapiStwa,
    CanadianSyllabicsNaskapiSkwa,
    CanadianSyllabicsNaskapiScwa,
    CanadianSyllabicsShe,
    CanadianSyllabicsShi,
    CanadianSyllabicsShii,
    CanadianSyllabicsSho,
    CanadianSyllabicsShoo,
    CanadianSyllabicsSha,
    CanadianSyllabicsShaa,
    CanadianSyllabicsShwe,
    CanadianSyllabicsWestDashCreeShwe,
    CanadianSyllabicsShwi,
    CanadianSyllabicsWestDashCreeShwi,
    CanadianSyllabicsShwii,
    CanadianSyllabicsWestDashCreeShwii,
    CanadianSyllabicsShwo,
    CanadianSyllabicsWestDashCreeShwo,
    CanadianSyllabicsShwoo,
    CanadianSyllabicsWestDashCreeShwoo,
    CanadianSyllabicsShwa,
    CanadianSyllabicsWestDashCreeShwa,
    CanadianSyllabicsShwaa,
    CanadianSyllabicsWestDashCreeShwaa,
    CanadianSyllabicsSh,
    CanadianSyllabicsYe,
    CanadianSyllabicsYaai,
    CanadianSyllabicsYi,
    CanadianSyllabicsYii,
    CanadianSyllabicsYo,
    CanadianSyllabicsYoo,
    CanadianSyllabicsYDashCreeYoo,
    CanadianSyllabicsYa,
    CanadianSyllabicsYaa,
    CanadianSyllabicsYwe,
    CanadianSyllabicsWestDashCreeYwe,
    CanadianSyllabicsYwi,
    CanadianSyllabicsWestDashCreeYwi,
    CanadianSyllabicsYwii,
    CanadianSyllabicsWestDashCreeYwii,
    CanadianSyllabicsYwo,
    CanadianSyllabicsWestDashCreeYwo,
    CanadianSyllabicsYwoo,
    CanadianSyllabicsWestDashCreeYwoo,
    CanadianSyllabicsYwa,
    CanadianSyllabicsWestDashCreeYwa,
    CanadianSyllabicsYwaa,
    CanadianSyllabicsWestDashCreeYwaa,
    CanadianSyllabicsNaskapiYwaa,
    CanadianSyllabicsY,
    CanadianSyllabicsBibleDashCreeY,
    CanadianSyllabicsWestDashCreeY,
    CanadianSyllabicsSayisiYi,
    CanadianSyllabicsRe,
    CanadianSyllabicsRDashCreeRe,
    CanadianSyllabicsWestDashCreeLe,
    CanadianSyllabicsRaai,
    CanadianSyllabicsRi,
    CanadianSyllabicsRii,
    CanadianSyllabicsRo,
    CanadianSyllabicsRoo,
    CanadianSyllabicsWestDashCreeLo,
    CanadianSyllabicsRa,
    CanadianSyllabicsRaa,
    CanadianSyllabicsWestDashCreeLa,
    CanadianSyllabicsRwaa,
    CanadianSyllabicsWestDashCreeRwaa,
    CanadianSyllabicsR,
    CanadianSyllabicsWestDashCreeR,
    CanadianSyllabicsMedialR,
    CanadianSyllabicsFe,
    CanadianSyllabicsFaai,
    CanadianSyllabicsFi,
    CanadianSyllabicsFii,
    CanadianSyllabicsFo,
    CanadianSyllabicsFoo,
    CanadianSyllabicsFa,
    CanadianSyllabicsFaa,
    CanadianSyllabicsFwaa,
    CanadianSyllabicsWestDashCreeFwaa,
    CanadianSyllabicsF,
    CanadianSyllabicsThe,
    CanadianSyllabicsNDashCreeThe,
    CanadianSyllabicsThi,
    CanadianSyllabicsNDashCreeThi,
    CanadianSyllabicsThii,
    CanadianSyllabicsNDashCreeThii,
    CanadianSyllabicsTho,
    CanadianSyllabicsThoo,
    CanadianSyllabicsTha,
    CanadianSyllabicsThaa,
    CanadianSyllabicsThwaa,
    CanadianSyllabicsWestDashCreeThwaa,
    CanadianSyllabicsTh,
    CanadianSyllabicsTthe,
    CanadianSyllabicsTthi,
    CanadianSyllabicsTtho,
    CanadianSyllabicsTtha,
    CanadianSyllabicsTth,
    CanadianSyllabicsTye,
    CanadianSyllabicsTyi,
    CanadianSyllabicsTyo,
    CanadianSyllabicsTya,
    CanadianSyllabicsNunavikHe,
    CanadianSyllabicsNunavikHi,
    CanadianSyllabicsNunavikHii,
    CanadianSyllabicsNunavikHo,
    CanadianSyllabicsNunavikHoo,
    CanadianSyllabicsNunavikHa,
    CanadianSyllabicsNunavikHaa,
    CanadianSyllabicsNunavikH,
    CanadianSyllabicsNunavutH,
    CanadianSyllabicsHk,
    CanadianSyllabicsQaai,
    CanadianSyllabicsQi,
    CanadianSyllabicsQii,
    CanadianSyllabicsQo,
    CanadianSyllabicsQoo,
    CanadianSyllabicsQa,
    CanadianSyllabicsQaa,
    CanadianSyllabicsQ,
    CanadianSyllabicsTlhe,
    CanadianSyllabicsTlhi,
    CanadianSyllabicsTlho,
    CanadianSyllabicsTlha,
    CanadianSyllabicsWestDashCreeRe,
    CanadianSyllabicsWestDashCreeRi,
    CanadianSyllabicsWestDashCreeRo,
    CanadianSyllabicsWestDashCreeRa,
    CanadianSyllabicsNgaai,
    CanadianSyllabicsNgi,
    CanadianSyllabicsNgii,
    CanadianSyllabicsNgo,
    CanadianSyllabicsNgoo,
    CanadianSyllabicsNga,
    CanadianSyllabicsNgaa,
    CanadianSyllabicsNg,
    CanadianSyllabicsNng,
    CanadianSyllabicsSayisiShe,
    CanadianSyllabicsSayisiShi,
    CanadianSyllabicsSayisiSho,
    CanadianSyllabicsSayisiSha,
    CanadianSyllabicsWoodsDashCreeThe,
    CanadianSyllabicsWoodsDashCreeThi,
    CanadianSyllabicsWoodsDashCreeTho,
    CanadianSyllabicsWoodsDashCreeTha,
    CanadianSyllabicsWoodsDashCreeTh,
    CanadianSyllabicsLhi,
    CanadianSyllabicsLhii,
    CanadianSyllabicsLho,
    CanadianSyllabicsLhoo,
    CanadianSyllabicsLha,
    CanadianSyllabicsLhaa,
    CanadianSyllabicsLh,
    CanadianSyllabicsThDashCreeThe,
    CanadianSyllabicsThDashCreeThi,
    CanadianSyllabicsThDashCreeThii,
    CanadianSyllabicsThDashCreeTho,
    CanadianSyllabicsThDashCreeThoo,
    CanadianSyllabicsThDashCreeTha,
    CanadianSyllabicsThDashCreeThaa,
    CanadianSyllabicsThDashCreeTh,
    CanadianSyllabicsAivilikB,
    CanadianSyllabicsBlackfootE,
    CanadianSyllabicsBlackfootI,
    CanadianSyllabicsBlackfootO,
    CanadianSyllabicsBlackfootA,
    CanadianSyllabicsBlackfootWe,
    CanadianSyllabicsBlackfootWi,
    CanadianSyllabicsBlackfootWo,
    CanadianSyllabicsBlackfootWa,
    CanadianSyllabicsBlackfootNe,
    CanadianSyllabicsBlackfootNi,
    CanadianSyllabicsBlackfootNo,
    CanadianSyllabicsBlackfootNa,
    CanadianSyllabicsBlackfootKe,
    CanadianSyllabicsBlackfootKi,
    CanadianSyllabicsBlackfootKo,
    CanadianSyllabicsBlackfootKa,
    CanadianSyllabicsSayisiHe,
    CanadianSyllabicsSayisiHi,
    CanadianSyllabicsSayisiHo,
    CanadianSyllabicsSayisiHa,
    CanadianSyllabicsCarrierGhu,
    CanadianSyllabicsCarrierGho,
    CanadianSyllabicsCarrierGhe,
    CanadianSyllabicsCarrierGhee,
    CanadianSyllabicsCarrierGhi,
    CanadianSyllabicsCarrierGha,
    CanadianSyllabicsCarrierRu,
    CanadianSyllabicsCarrierRo,
    CanadianSyllabicsCarrierRe,
    CanadianSyllabicsCarrierRee,
    CanadianSyllabicsCarrierRi,
    CanadianSyllabicsCarrierRa,
    CanadianSyllabicsCarrierWu,
    CanadianSyllabicsCarrierWo,
    CanadianSyllabicsCarrierWe,
    CanadianSyllabicsCarrierWee,
    CanadianSyllabicsCarrierWi,
    CanadianSyllabicsCarrierWa,
    CanadianSyllabicsCarrierHwu,
    CanadianSyllabicsCarrierHwo,
    CanadianSyllabicsCarrierHwe,
    CanadianSyllabicsCarrierHwee,
    CanadianSyllabicsCarrierHwi,
    CanadianSyllabicsCarrierHwa,
    CanadianSyllabicsCarrierThu,
    CanadianSyllabicsCarrierTho,
    CanadianSyllabicsCarrierThe,
    CanadianSyllabicsCarrierThee,
    CanadianSyllabicsCarrierThi,
    CanadianSyllabicsCarrierTha,
    CanadianSyllabicsCarrierTtu,
    CanadianSyllabicsCarrierTto,
    CanadianSyllabicsCarrierTte,
    CanadianSyllabicsCarrierTtee,
    CanadianSyllabicsCarrierTti,
    CanadianSyllabicsCarrierTta,
    CanadianSyllabicsCarrierPu,
    CanadianSyllabicsCarrierPo,
    CanadianSyllabicsCarrierPe,
    CanadianSyllabicsCarrierPee,
    CanadianSyllabicsCarrierPi,
    CanadianSyllabicsCarrierPa,
    CanadianSyllabicsCarrierP,
    CanadianSyllabicsCarrierGu,
    CanadianSyllabicsCarrierGo,
    CanadianSyllabicsCarrierGe,
    CanadianSyllabicsCarrierGee,
    CanadianSyllabicsCarrierGi,
    CanadianSyllabicsCarrierGa,
    CanadianSyllabicsCarrierKhu,
    CanadianSyllabicsCarrierKho,
    CanadianSyllabicsCarrierKhe,
    CanadianSyllabicsCarrierKhee,
    CanadianSyllabicsCarrierKhi,
    CanadianSyllabicsCarrierKha,
    CanadianSyllabicsCarrierKku,
    CanadianSyllabicsCarrierKko,
    CanadianSyllabicsCarrierKke,
    CanadianSyllabicsCarrierKkee,
    CanadianSyllabicsCarrierKki,
    CanadianSyllabicsCarrierKka,
    CanadianSyllabicsCarrierKk,
    CanadianSyllabicsCarrierNu,
    CanadianSyllabicsCarrierNo,
    CanadianSyllabicsCarrierNe,
    CanadianSyllabicsCarrierNee,
    CanadianSyllabicsCarrierNi,
    CanadianSyllabicsCarrierNa,
    CanadianSyllabicsCarrierMu,
    CanadianSyllabicsCarrierMo,
    CanadianSyllabicsCarrierMe,
    CanadianSyllabicsCarrierMee,
    CanadianSyllabicsCarrierMi,
    CanadianSyllabicsCarrierMa,
    CanadianSyllabicsCarrierYu,
    CanadianSyllabicsCarrierYo,
    CanadianSyllabicsCarrierYe,
    CanadianSyllabicsCarrierYee,
    CanadianSyllabicsCarrierYi,
    CanadianSyllabicsCarrierYa,
    CanadianSyllabicsCarrierJu,
    CanadianSyllabicsSayisiJu,
    CanadianSyllabicsCarrierJo,
    CanadianSyllabicsCarrierJe,
    CanadianSyllabicsCarrierJee,
    CanadianSyllabicsCarrierJi,
    CanadianSyllabicsSayisiJi,
    CanadianSyllabicsCarrierJa,
    CanadianSyllabicsCarrierJju,
    CanadianSyllabicsCarrierJjo,
    CanadianSyllabicsCarrierJje,
    CanadianSyllabicsCarrierJjee,
    CanadianSyllabicsCarrierJji,
    CanadianSyllabicsCarrierJja,
    CanadianSyllabicsCarrierLu,
    CanadianSyllabicsCarrierLo,
    CanadianSyllabicsCarrierLe,
    CanadianSyllabicsCarrierLee,
    CanadianSyllabicsCarrierLi,
    CanadianSyllabicsCarrierLa,
    CanadianSyllabicsCarrierDlu,
    CanadianSyllabicsCarrierDlo,
    CanadianSyllabicsCarrierDle,
    CanadianSyllabicsCarrierDlee,
    CanadianSyllabicsCarrierDli,
    CanadianSyllabicsCarrierDla,
    CanadianSyllabicsCarrierLhu,
    CanadianSyllabicsCarrierLho,
    CanadianSyllabicsCarrierLhe,
    CanadianSyllabicsCarrierLhee,
    CanadianSyllabicsCarrierLhi,
    CanadianSyllabicsCarrierLha,
    CanadianSyllabicsCarrierTlhu,
    CanadianSyllabicsCarrierTlho,
    CanadianSyllabicsCarrierTlhe,
    CanadianSyllabicsCarrierTlhee,
    CanadianSyllabicsCarrierTlhi,
    CanadianSyllabicsCarrierTlha,
    CanadianSyllabicsCarrierTlu,
    CanadianSyllabicsCarrierTlo,
    CanadianSyllabicsCarrierTle,
    CanadianSyllabicsCarrierTlee,
    CanadianSyllabicsCarrierTli,
    CanadianSyllabicsCarrierTla,
    CanadianSyllabicsCarrierZu,
    CanadianSyllabicsCarrierZo,
    CanadianSyllabicsCarrierZe,
    CanadianSyllabicsCarrierZee,
    CanadianSyllabicsCarrierZi,
    CanadianSyllabicsCarrierZa,
    CanadianSyllabicsCarrierZ,
    CanadianSyllabicsCarrierInitialZ,
    CanadianSyllabicsCarrierDzu,
    CanadianSyllabicsCarrierDzo,
    CanadianSyllabicsCarrierDze,
    CanadianSyllabicsCarrierDzee,
    CanadianSyllabicsCarrierDzi,
    CanadianSyllabicsCarrierDza,
    CanadianSyllabicsCarrierSu,
    CanadianSyllabicsCarrierSo,
    CanadianSyllabicsCarrierSe,
    CanadianSyllabicsCarrierSee,
    CanadianSyllabicsCarrierSi,
    CanadianSyllabicsCarrierSa,
    CanadianSyllabicsCarrierShu,
    CanadianSyllabicsCarrierSho,
    CanadianSyllabicsCarrierShe,
    CanadianSyllabicsCarrierShee,
    CanadianSyllabicsCarrierShi,
    CanadianSyllabicsCarrierSha,
    CanadianSyllabicsCarrierSh,
    CanadianSyllabicsCarrierTsu,
    CanadianSyllabicsCarrierTso,
    CanadianSyllabicsCarrierTse,
    CanadianSyllabicsCarrierTsee,
    CanadianSyllabicsCarrierTsi,
    CanadianSyllabicsCarrierTsa,
    CanadianSyllabicsCarrierChu,
    CanadianSyllabicsCarrierCho,
    CanadianSyllabicsCarrierChe,
    CanadianSyllabicsCarrierChee,
    CanadianSyllabicsCarrierChi,
    CanadianSyllabicsCarrierCha,
    CanadianSyllabicsCarrierTtsu,
    CanadianSyllabicsCarrierTtso,
    CanadianSyllabicsCarrierTtse,
    CanadianSyllabicsCarrierTtsee,
    CanadianSyllabicsCarrierTtsi,
    CanadianSyllabicsCarrierTtsa,
    CanadianSyllabicsChiSign,
    CanadianSyllabicsFullStop,
    CanadianSyllabicsQai,
    CanadianSyllabicsNgai,
    CanadianSyllabicsNngi,
    CanadianSyllabicsNngii,
    CanadianSyllabicsNngo,
    CanadianSyllabicsNngoo,
    CanadianSyllabicsNnga,
    CanadianSyllabicsNngaa,
    CanadianSyllabicsWoodsDashCreeThwee,
    CanadianSyllabicsWoodsDashCreeThwi,
    CanadianSyllabicsWoodsDashCreeThwii,
    CanadianSyllabicsWoodsDashCreeThwo,
    CanadianSyllabicsWoodsDashCreeThwoo,
    CanadianSyllabicsWoodsDashCreeThwa,
    CanadianSyllabicsWoodsDashCreeThwaa,
    CanadianSyllabicsWoodsDashCreeFinalTh,
}

An enum to represent all characters in the UnifiedCanadianAboriginalSyllabics block.

Variants
CanadianSyllabicsHyphen

\u{1400}: ''

CanadianSyllabicsE

\u{1401}: ''

CanadianSyllabicsAai

\u{1402}: ''

CanadianSyllabicsI

\u{1403}: ''

CanadianSyllabicsIi

\u{1404}: ''

CanadianSyllabicsO

\u{1405}: ''

CanadianSyllabicsOo

\u{1406}: ''

CanadianSyllabicsYDashCreeOo

\u{1407}: ''

CanadianSyllabicsCarrierEe

\u{1408}: ''

CanadianSyllabicsCarrierI

\u{1409}: ''

CanadianSyllabicsA

\u{140a}: ''

CanadianSyllabicsAa

\u{140b}: ''

CanadianSyllabicsWe

\u{140c}: ''

CanadianSyllabicsWestDashCreeWe

\u{140d}: ''

CanadianSyllabicsWi

\u{140e}: ''

CanadianSyllabicsWestDashCreeWi

\u{140f}: ''

CanadianSyllabicsWii

\u{1410}: ''

CanadianSyllabicsWestDashCreeWii

\u{1411}: ''

CanadianSyllabicsWo

\u{1412}: ''

CanadianSyllabicsWestDashCreeWo

\u{1413}: ''

CanadianSyllabicsWoo

\u{1414}: ''

CanadianSyllabicsWestDashCreeWoo

\u{1415}: ''

CanadianSyllabicsNaskapiWoo

\u{1416}: ''

CanadianSyllabicsWa

\u{1417}: ''

CanadianSyllabicsWestDashCreeWa

\u{1418}: ''

CanadianSyllabicsWaa

\u{1419}: ''

CanadianSyllabicsWestDashCreeWaa

\u{141a}: ''

CanadianSyllabicsNaskapiWaa

\u{141b}: ''

CanadianSyllabicsAi

\u{141c}: ''

CanadianSyllabicsYDashCreeW

\u{141d}: ''

CanadianSyllabicsGlottalStop

\u{141e}: ''

CanadianSyllabicsFinalAcute

\u{141f}: ''

CanadianSyllabicsFinalGrave

\u{1420}: ''

CanadianSyllabicsFinalBottomHalfRing

\u{1421}: ''

CanadianSyllabicsFinalTopHalfRing

\u{1422}: ''

CanadianSyllabicsFinalRightHalfRing

\u{1423}: ''

CanadianSyllabicsFinalRing

\u{1424}: ''

CanadianSyllabicsFinalDoubleAcute

\u{1425}: ''

CanadianSyllabicsFinalDoubleShortVerticalStrokes

\u{1426}: ''

CanadianSyllabicsFinalMiddleDot

\u{1427}: ''

CanadianSyllabicsFinalShortHorizontalStroke

\u{1428}: ''

CanadianSyllabicsFinalPlus

\u{1429}: ''

CanadianSyllabicsFinalDownTack

\u{142a}: ''

CanadianSyllabicsEn

\u{142b}: ''

CanadianSyllabicsIn

\u{142c}: ''

CanadianSyllabicsOn

\u{142d}: ''

CanadianSyllabicsAn

\u{142e}: ''

CanadianSyllabicsPe

\u{142f}: ''

CanadianSyllabicsPaai

\u{1430}: ''

CanadianSyllabicsPi

\u{1431}: ''

CanadianSyllabicsPii

\u{1432}: ''

CanadianSyllabicsPo

\u{1433}: ''

CanadianSyllabicsPoo

\u{1434}: ''

CanadianSyllabicsYDashCreePoo

\u{1435}: ''

CanadianSyllabicsCarrierHee

\u{1436}: ''

CanadianSyllabicsCarrierHi

\u{1437}: ''

CanadianSyllabicsPa

\u{1438}: ''

CanadianSyllabicsPaa

\u{1439}: ''

CanadianSyllabicsPwe

\u{143a}: ''

CanadianSyllabicsWestDashCreePwe

\u{143b}: ''

CanadianSyllabicsPwi

\u{143c}: ''

CanadianSyllabicsWestDashCreePwi

\u{143d}: ''

CanadianSyllabicsPwii

\u{143e}: ''

CanadianSyllabicsWestDashCreePwii

\u{143f}: ''

CanadianSyllabicsPwo

\u{1440}: ''

CanadianSyllabicsWestDashCreePwo

\u{1441}: ''

CanadianSyllabicsPwoo

\u{1442}: ''

CanadianSyllabicsWestDashCreePwoo

\u{1443}: ''

CanadianSyllabicsPwa

\u{1444}: ''

CanadianSyllabicsWestDashCreePwa

\u{1445}: ''

CanadianSyllabicsPwaa

\u{1446}: ''

CanadianSyllabicsWestDashCreePwaa

\u{1447}: ''

CanadianSyllabicsYDashCreePwaa

\u{1448}: ''

CanadianSyllabicsP

\u{1449}: ''

CanadianSyllabicsWestDashCreeP

\u{144a}: ''

CanadianSyllabicsCarrierH

\u{144b}: ''

CanadianSyllabicsTe

\u{144c}: ''

CanadianSyllabicsTaai

\u{144d}: ''

CanadianSyllabicsTi

\u{144e}: ''

CanadianSyllabicsTii

\u{144f}: ''

CanadianSyllabicsTo

\u{1450}: ''

CanadianSyllabicsToo

\u{1451}: ''

CanadianSyllabicsYDashCreeToo

\u{1452}: ''

CanadianSyllabicsCarrierDee

\u{1453}: ''

CanadianSyllabicsCarrierDi

\u{1454}: ''

CanadianSyllabicsTa

\u{1455}: ''

CanadianSyllabicsTaa

\u{1456}: ''

CanadianSyllabicsTwe

\u{1457}: ''

CanadianSyllabicsWestDashCreeTwe

\u{1458}: ''

CanadianSyllabicsTwi

\u{1459}: ''

CanadianSyllabicsWestDashCreeTwi

\u{145a}: ''

CanadianSyllabicsTwii

\u{145b}: ''

CanadianSyllabicsWestDashCreeTwii

\u{145c}: ''

CanadianSyllabicsTwo

\u{145d}: ''

CanadianSyllabicsWestDashCreeTwo

\u{145e}: ''

CanadianSyllabicsTwoo

\u{145f}: ''

CanadianSyllabicsWestDashCreeTwoo

\u{1460}: ''

CanadianSyllabicsTwa

\u{1461}: ''

CanadianSyllabicsWestDashCreeTwa

\u{1462}: ''

CanadianSyllabicsTwaa

\u{1463}: ''

CanadianSyllabicsWestDashCreeTwaa

\u{1464}: ''

CanadianSyllabicsNaskapiTwaa

\u{1465}: ''

CanadianSyllabicsT

\u{1466}: ''

CanadianSyllabicsTte

\u{1467}: ''

CanadianSyllabicsTti

\u{1468}: ''

CanadianSyllabicsTto

\u{1469}: ''

CanadianSyllabicsTta

\u{146a}: ''

CanadianSyllabicsKe

\u{146b}: ''

CanadianSyllabicsKaai

\u{146c}: ''

CanadianSyllabicsKi

\u{146d}: ''

CanadianSyllabicsKii

\u{146e}: ''

CanadianSyllabicsKo

\u{146f}: ''

CanadianSyllabicsKoo

\u{1470}: ''

CanadianSyllabicsYDashCreeKoo

\u{1471}: ''

CanadianSyllabicsKa

\u{1472}: ''

CanadianSyllabicsKaa

\u{1473}: ''

CanadianSyllabicsKwe

\u{1474}: ''

CanadianSyllabicsWestDashCreeKwe

\u{1475}: ''

CanadianSyllabicsKwi

\u{1476}: ''

CanadianSyllabicsWestDashCreeKwi

\u{1477}: ''

CanadianSyllabicsKwii

\u{1478}: ''

CanadianSyllabicsWestDashCreeKwii

\u{1479}: ''

CanadianSyllabicsKwo

\u{147a}: ''

CanadianSyllabicsWestDashCreeKwo

\u{147b}: ''

CanadianSyllabicsKwoo

\u{147c}: ''

CanadianSyllabicsWestDashCreeKwoo

\u{147d}: ''

CanadianSyllabicsKwa

\u{147e}: ''

CanadianSyllabicsWestDashCreeKwa

\u{147f}: ''

CanadianSyllabicsKwaa

\u{1480}: ''

CanadianSyllabicsWestDashCreeKwaa

\u{1481}: ''

CanadianSyllabicsNaskapiKwaa

\u{1482}: ''

CanadianSyllabicsK

\u{1483}: ''

CanadianSyllabicsKw

\u{1484}: ''

CanadianSyllabicsSouthDashSlaveyKeh

\u{1485}: ''

CanadianSyllabicsSouthDashSlaveyKih

\u{1486}: ''

CanadianSyllabicsSouthDashSlaveyKoh

\u{1487}: ''

CanadianSyllabicsSouthDashSlaveyKah

\u{1488}: ''

CanadianSyllabicsCe

\u{1489}: ''

CanadianSyllabicsCaai

\u{148a}: ''

CanadianSyllabicsCi

\u{148b}: ''

CanadianSyllabicsCii

\u{148c}: ''

CanadianSyllabicsCo

\u{148d}: ''

CanadianSyllabicsCoo

\u{148e}: ''

CanadianSyllabicsYDashCreeCoo

\u{148f}: ''

CanadianSyllabicsCa

\u{1490}: ''

CanadianSyllabicsCaa

\u{1491}: ''

CanadianSyllabicsCwe

\u{1492}: ''

CanadianSyllabicsWestDashCreeCwe

\u{1493}: ''

CanadianSyllabicsCwi

\u{1494}: ''

CanadianSyllabicsWestDashCreeCwi

\u{1495}: ''

CanadianSyllabicsCwii

\u{1496}: ''

CanadianSyllabicsWestDashCreeCwii

\u{1497}: ''

CanadianSyllabicsCwo

\u{1498}: ''

CanadianSyllabicsWestDashCreeCwo

\u{1499}: ''

CanadianSyllabicsCwoo

\u{149a}: ''

CanadianSyllabicsWestDashCreeCwoo

\u{149b}: ''

CanadianSyllabicsCwa

\u{149c}: ''

CanadianSyllabicsWestDashCreeCwa

\u{149d}: ''

CanadianSyllabicsCwaa

\u{149e}: ''

CanadianSyllabicsWestDashCreeCwaa

\u{149f}: ''

CanadianSyllabicsNaskapiCwaa

\u{14a0}: ''

CanadianSyllabicsC

\u{14a1}: ''

CanadianSyllabicsSayisiTh

\u{14a2}: ''

CanadianSyllabicsMe

\u{14a3}: ''

CanadianSyllabicsMaai

\u{14a4}: ''

CanadianSyllabicsMi

\u{14a5}: ''

CanadianSyllabicsMii

\u{14a6}: ''

CanadianSyllabicsMo

\u{14a7}: ''

CanadianSyllabicsMoo

\u{14a8}: ''

CanadianSyllabicsYDashCreeMoo

\u{14a9}: ''

CanadianSyllabicsMa

\u{14aa}: ''

CanadianSyllabicsMaa

\u{14ab}: ''

CanadianSyllabicsMwe

\u{14ac}: ''

CanadianSyllabicsWestDashCreeMwe

\u{14ad}: ''

CanadianSyllabicsMwi

\u{14ae}: ''

CanadianSyllabicsWestDashCreeMwi

\u{14af}: ''

CanadianSyllabicsMwii

\u{14b0}: ''

CanadianSyllabicsWestDashCreeMwii

\u{14b1}: ''

CanadianSyllabicsMwo

\u{14b2}: ''

CanadianSyllabicsWestDashCreeMwo

\u{14b3}: ''

CanadianSyllabicsMwoo

\u{14b4}: ''

CanadianSyllabicsWestDashCreeMwoo

\u{14b5}: ''

CanadianSyllabicsMwa

\u{14b6}: ''

CanadianSyllabicsWestDashCreeMwa

\u{14b7}: ''

CanadianSyllabicsMwaa

\u{14b8}: ''

CanadianSyllabicsWestDashCreeMwaa

\u{14b9}: ''

CanadianSyllabicsNaskapiMwaa

\u{14ba}: ''

CanadianSyllabicsM

\u{14bb}: ''

CanadianSyllabicsWestDashCreeM

\u{14bc}: ''

CanadianSyllabicsMh

\u{14bd}: ''

CanadianSyllabicsAthapascanM

\u{14be}: ''

CanadianSyllabicsSayisiM

\u{14bf}: ''

CanadianSyllabicsNe

\u{14c0}: ''

CanadianSyllabicsNaai

\u{14c1}: ''

CanadianSyllabicsNi

\u{14c2}: ''

CanadianSyllabicsNii

\u{14c3}: ''

CanadianSyllabicsNo

\u{14c4}: ''

CanadianSyllabicsNoo

\u{14c5}: ''

CanadianSyllabicsYDashCreeNoo

\u{14c6}: ''

CanadianSyllabicsNa

\u{14c7}: ''

CanadianSyllabicsNaa

\u{14c8}: ''

CanadianSyllabicsNwe

\u{14c9}: ''

CanadianSyllabicsWestDashCreeNwe

\u{14ca}: ''

CanadianSyllabicsNwa

\u{14cb}: ''

CanadianSyllabicsWestDashCreeNwa

\u{14cc}: ''

CanadianSyllabicsNwaa

\u{14cd}: ''

CanadianSyllabicsWestDashCreeNwaa

\u{14ce}: ''

CanadianSyllabicsNaskapiNwaa

\u{14cf}: ''

CanadianSyllabicsN

\u{14d0}: ''

CanadianSyllabicsCarrierNg

\u{14d1}: ''

CanadianSyllabicsNh

\u{14d2}: ''

CanadianSyllabicsLe

\u{14d3}: ''

CanadianSyllabicsLaai

\u{14d4}: ''

CanadianSyllabicsLi

\u{14d5}: ''

CanadianSyllabicsLii

\u{14d6}: ''

CanadianSyllabicsLo

\u{14d7}: ''

CanadianSyllabicsLoo

\u{14d8}: ''

CanadianSyllabicsYDashCreeLoo

\u{14d9}: ''

CanadianSyllabicsLa

\u{14da}: ''

CanadianSyllabicsLaa

\u{14db}: ''

CanadianSyllabicsLwe

\u{14dc}: ''

CanadianSyllabicsWestDashCreeLwe

\u{14dd}: ''

CanadianSyllabicsLwi

\u{14de}: ''

CanadianSyllabicsWestDashCreeLwi

\u{14df}: ''

CanadianSyllabicsLwii

\u{14e0}: ''

CanadianSyllabicsWestDashCreeLwii

\u{14e1}: ''

CanadianSyllabicsLwo

\u{14e2}: ''

CanadianSyllabicsWestDashCreeLwo

\u{14e3}: ''

CanadianSyllabicsLwoo

\u{14e4}: ''

CanadianSyllabicsWestDashCreeLwoo

\u{14e5}: ''

CanadianSyllabicsLwa

\u{14e6}: ''

CanadianSyllabicsWestDashCreeLwa

\u{14e7}: ''

CanadianSyllabicsLwaa

\u{14e8}: ''

CanadianSyllabicsWestDashCreeLwaa

\u{14e9}: ''

CanadianSyllabicsL

\u{14ea}: ''

CanadianSyllabicsWestDashCreeL

\u{14eb}: ''

CanadianSyllabicsMedialL

\u{14ec}: ''

CanadianSyllabicsSe

\u{14ed}: ''

CanadianSyllabicsSaai

\u{14ee}: ''

CanadianSyllabicsSi

\u{14ef}: ''

CanadianSyllabicsSii

\u{14f0}: ''

CanadianSyllabicsSo

\u{14f1}: ''

CanadianSyllabicsSoo

\u{14f2}: ''

CanadianSyllabicsYDashCreeSoo

\u{14f3}: ''

CanadianSyllabicsSa

\u{14f4}: ''

CanadianSyllabicsSaa

\u{14f5}: ''

CanadianSyllabicsSwe

\u{14f6}: ''

CanadianSyllabicsWestDashCreeSwe

\u{14f7}: ''

CanadianSyllabicsSwi

\u{14f8}: ''

CanadianSyllabicsWestDashCreeSwi

\u{14f9}: ''

CanadianSyllabicsSwii

\u{14fa}: ''

CanadianSyllabicsWestDashCreeSwii

\u{14fb}: ''

CanadianSyllabicsSwo

\u{14fc}: ''

CanadianSyllabicsWestDashCreeSwo

\u{14fd}: ''

CanadianSyllabicsSwoo

\u{14fe}: ''

CanadianSyllabicsWestDashCreeSwoo

\u{14ff}: ''

CanadianSyllabicsSwa

\u{1500}: ''

CanadianSyllabicsWestDashCreeSwa

\u{1501}: ''

CanadianSyllabicsSwaa

\u{1502}: ''

CanadianSyllabicsWestDashCreeSwaa

\u{1503}: ''

CanadianSyllabicsNaskapiSwaa

\u{1504}: ''

CanadianSyllabicsS

\u{1505}: ''

CanadianSyllabicsAthapascanS

\u{1506}: ''

CanadianSyllabicsSw

\u{1507}: ''

CanadianSyllabicsBlackfootS

\u{1508}: ''

CanadianSyllabicsMooseDashCreeSk

\u{1509}: ''

CanadianSyllabicsNaskapiSkw

\u{150a}: ''

CanadianSyllabicsNaskapiSDashW

\u{150b}: ''

CanadianSyllabicsNaskapiSpwa

\u{150c}: ''

CanadianSyllabicsNaskapiStwa

\u{150d}: ''

CanadianSyllabicsNaskapiSkwa

\u{150e}: ''

CanadianSyllabicsNaskapiScwa

\u{150f}: ''

CanadianSyllabicsShe

\u{1510}: ''

CanadianSyllabicsShi

\u{1511}: ''

CanadianSyllabicsShii

\u{1512}: ''

CanadianSyllabicsSho

\u{1513}: ''

CanadianSyllabicsShoo

\u{1514}: ''

CanadianSyllabicsSha

\u{1515}: ''

CanadianSyllabicsShaa

\u{1516}: ''

CanadianSyllabicsShwe

\u{1517}: ''

CanadianSyllabicsWestDashCreeShwe

\u{1518}: ''

CanadianSyllabicsShwi

\u{1519}: ''

CanadianSyllabicsWestDashCreeShwi

\u{151a}: ''

CanadianSyllabicsShwii

\u{151b}: ''

CanadianSyllabicsWestDashCreeShwii

\u{151c}: ''

CanadianSyllabicsShwo

\u{151d}: ''

CanadianSyllabicsWestDashCreeShwo

\u{151e}: ''

CanadianSyllabicsShwoo

\u{151f}: ''

CanadianSyllabicsWestDashCreeShwoo

\u{1520}: ''

CanadianSyllabicsShwa

\u{1521}: ''

CanadianSyllabicsWestDashCreeShwa

\u{1522}: ''

CanadianSyllabicsShwaa

\u{1523}: ''

CanadianSyllabicsWestDashCreeShwaa

\u{1524}: ''

CanadianSyllabicsSh

\u{1525}: ''

CanadianSyllabicsYe

\u{1526}: ''

CanadianSyllabicsYaai

\u{1527}: ''

CanadianSyllabicsYi

\u{1528}: ''

CanadianSyllabicsYii

\u{1529}: ''

CanadianSyllabicsYo

\u{152a}: ''

CanadianSyllabicsYoo

\u{152b}: ''

CanadianSyllabicsYDashCreeYoo

\u{152c}: ''

CanadianSyllabicsYa

\u{152d}: ''

CanadianSyllabicsYaa

\u{152e}: ''

CanadianSyllabicsYwe

\u{152f}: ''

CanadianSyllabicsWestDashCreeYwe

\u{1530}: ''

CanadianSyllabicsYwi

\u{1531}: ''

CanadianSyllabicsWestDashCreeYwi

\u{1532}: ''

CanadianSyllabicsYwii

\u{1533}: ''

CanadianSyllabicsWestDashCreeYwii

\u{1534}: ''

CanadianSyllabicsYwo

\u{1535}: ''

CanadianSyllabicsWestDashCreeYwo

\u{1536}: ''

CanadianSyllabicsYwoo

\u{1537}: ''

CanadianSyllabicsWestDashCreeYwoo

\u{1538}: ''

CanadianSyllabicsYwa

\u{1539}: ''

CanadianSyllabicsWestDashCreeYwa

\u{153a}: ''

CanadianSyllabicsYwaa

\u{153b}: ''

CanadianSyllabicsWestDashCreeYwaa

\u{153c}: ''

CanadianSyllabicsNaskapiYwaa

\u{153d}: ''

CanadianSyllabicsY

\u{153e}: ''

CanadianSyllabicsBibleDashCreeY

\u{153f}: ''

CanadianSyllabicsWestDashCreeY

\u{1540}: ''

CanadianSyllabicsSayisiYi

\u{1541}: ''

CanadianSyllabicsRe

\u{1542}: ''

CanadianSyllabicsRDashCreeRe

\u{1543}: ''

CanadianSyllabicsWestDashCreeLe

\u{1544}: ''

CanadianSyllabicsRaai

\u{1545}: ''

CanadianSyllabicsRi

\u{1546}: ''

CanadianSyllabicsRii

\u{1547}: ''

CanadianSyllabicsRo

\u{1548}: ''

CanadianSyllabicsRoo

\u{1549}: ''

CanadianSyllabicsWestDashCreeLo

\u{154a}: ''

CanadianSyllabicsRa

\u{154b}: ''

CanadianSyllabicsRaa

\u{154c}: ''

CanadianSyllabicsWestDashCreeLa

\u{154d}: ''

CanadianSyllabicsRwaa

\u{154e}: ''

CanadianSyllabicsWestDashCreeRwaa

\u{154f}: ''

CanadianSyllabicsR

\u{1550}: ''

CanadianSyllabicsWestDashCreeR

\u{1551}: ''

CanadianSyllabicsMedialR

\u{1552}: ''

CanadianSyllabicsFe

\u{1553}: ''

CanadianSyllabicsFaai

\u{1554}: ''

CanadianSyllabicsFi

\u{1555}: ''

CanadianSyllabicsFii

\u{1556}: ''

CanadianSyllabicsFo

\u{1557}: ''

CanadianSyllabicsFoo

\u{1558}: ''

CanadianSyllabicsFa

\u{1559}: ''

CanadianSyllabicsFaa

\u{155a}: ''

CanadianSyllabicsFwaa

\u{155b}: ''

CanadianSyllabicsWestDashCreeFwaa

\u{155c}: ''

CanadianSyllabicsF

\u{155d}: ''

CanadianSyllabicsThe

\u{155e}: ''

CanadianSyllabicsNDashCreeThe

\u{155f}: ''

CanadianSyllabicsThi

\u{1560}: ''

CanadianSyllabicsNDashCreeThi

\u{1561}: ''

CanadianSyllabicsThii

\u{1562}: ''

CanadianSyllabicsNDashCreeThii

\u{1563}: ''

CanadianSyllabicsTho

\u{1564}: ''

CanadianSyllabicsThoo

\u{1565}: ''

CanadianSyllabicsTha

\u{1566}: ''

CanadianSyllabicsThaa

\u{1567}: ''

CanadianSyllabicsThwaa

\u{1568}: ''

CanadianSyllabicsWestDashCreeThwaa

\u{1569}: ''

CanadianSyllabicsTh

\u{156a}: ''

CanadianSyllabicsTthe

\u{156b}: ''

CanadianSyllabicsTthi

\u{156c}: ''

CanadianSyllabicsTtho

\u{156d}: ''

CanadianSyllabicsTtha

\u{156e}: ''

CanadianSyllabicsTth

\u{156f}: ''

CanadianSyllabicsTye

\u{1570}: ''

CanadianSyllabicsTyi

\u{1571}: ''

CanadianSyllabicsTyo

\u{1572}: ''

CanadianSyllabicsTya

\u{1573}: ''

CanadianSyllabicsNunavikHe

\u{1574}: ''

CanadianSyllabicsNunavikHi

\u{1575}: ''

CanadianSyllabicsNunavikHii

\u{1576}: ''

CanadianSyllabicsNunavikHo

\u{1577}: ''

CanadianSyllabicsNunavikHoo

\u{1578}: ''

CanadianSyllabicsNunavikHa

\u{1579}: ''

CanadianSyllabicsNunavikHaa

\u{157a}: ''

CanadianSyllabicsNunavikH

\u{157b}: ''

CanadianSyllabicsNunavutH

\u{157c}: ''

CanadianSyllabicsHk

\u{157d}: ''

CanadianSyllabicsQaai

\u{157e}: ''

CanadianSyllabicsQi

\u{157f}: ''

CanadianSyllabicsQii

\u{1580}: ''

CanadianSyllabicsQo

\u{1581}: ''

CanadianSyllabicsQoo

\u{1582}: ''

CanadianSyllabicsQa

\u{1583}: ''

CanadianSyllabicsQaa

\u{1584}: ''

CanadianSyllabicsQ

\u{1585}: ''

CanadianSyllabicsTlhe

\u{1586}: ''

CanadianSyllabicsTlhi

\u{1587}: ''

CanadianSyllabicsTlho

\u{1588}: ''

CanadianSyllabicsTlha

\u{1589}: ''

CanadianSyllabicsWestDashCreeRe

\u{158a}: ''

CanadianSyllabicsWestDashCreeRi

\u{158b}: ''

CanadianSyllabicsWestDashCreeRo

\u{158c}: ''

CanadianSyllabicsWestDashCreeRa

\u{158d}: ''

CanadianSyllabicsNgaai

\u{158e}: ''

CanadianSyllabicsNgi

\u{158f}: ''

CanadianSyllabicsNgii

\u{1590}: ''

CanadianSyllabicsNgo

\u{1591}: ''

CanadianSyllabicsNgoo

\u{1592}: ''

CanadianSyllabicsNga

\u{1593}: ''

CanadianSyllabicsNgaa

\u{1594}: ''

CanadianSyllabicsNg

\u{1595}: ''

CanadianSyllabicsNng

\u{1596}: ''

CanadianSyllabicsSayisiShe

\u{1597}: ''

CanadianSyllabicsSayisiShi

\u{1598}: ''

CanadianSyllabicsSayisiSho

\u{1599}: ''

CanadianSyllabicsSayisiSha

\u{159a}: ''

CanadianSyllabicsWoodsDashCreeThe

\u{159b}: ''

CanadianSyllabicsWoodsDashCreeThi

\u{159c}: ''

CanadianSyllabicsWoodsDashCreeTho

\u{159d}: ''

CanadianSyllabicsWoodsDashCreeTha

\u{159e}: ''

CanadianSyllabicsWoodsDashCreeTh

\u{159f}: ''

CanadianSyllabicsLhi

\u{15a0}: ''

CanadianSyllabicsLhii

\u{15a1}: ''

CanadianSyllabicsLho

\u{15a2}: ''

CanadianSyllabicsLhoo

\u{15a3}: ''

CanadianSyllabicsLha

\u{15a4}: ''

CanadianSyllabicsLhaa

\u{15a5}: ''

CanadianSyllabicsLh

\u{15a6}: ''

CanadianSyllabicsThDashCreeThe

\u{15a7}: ''

CanadianSyllabicsThDashCreeThi

\u{15a8}: ''

CanadianSyllabicsThDashCreeThii

\u{15a9}: ''

CanadianSyllabicsThDashCreeTho

\u{15aa}: ''

CanadianSyllabicsThDashCreeThoo

\u{15ab}: ''

CanadianSyllabicsThDashCreeTha

\u{15ac}: ''

CanadianSyllabicsThDashCreeThaa

\u{15ad}: ''

CanadianSyllabicsThDashCreeTh

\u{15ae}: ''

CanadianSyllabicsAivilikB

\u{15af}: ''

CanadianSyllabicsBlackfootE

\u{15b0}: ''

CanadianSyllabicsBlackfootI

\u{15b1}: ''

CanadianSyllabicsBlackfootO

\u{15b2}: ''

CanadianSyllabicsBlackfootA

\u{15b3}: ''

CanadianSyllabicsBlackfootWe

\u{15b4}: ''

CanadianSyllabicsBlackfootWi

\u{15b5}: ''

CanadianSyllabicsBlackfootWo

\u{15b6}: ''

CanadianSyllabicsBlackfootWa

\u{15b7}: ''

CanadianSyllabicsBlackfootNe

\u{15b8}: ''

CanadianSyllabicsBlackfootNi

\u{15b9}: ''

CanadianSyllabicsBlackfootNo

\u{15ba}: ''

CanadianSyllabicsBlackfootNa

\u{15bb}: ''

CanadianSyllabicsBlackfootKe

\u{15bc}: ''

CanadianSyllabicsBlackfootKi

\u{15bd}: ''

CanadianSyllabicsBlackfootKo

\u{15be}: ''

CanadianSyllabicsBlackfootKa

\u{15bf}: ''

CanadianSyllabicsSayisiHe

\u{15c0}: ''

CanadianSyllabicsSayisiHi

\u{15c1}: ''

CanadianSyllabicsSayisiHo

\u{15c2}: ''

CanadianSyllabicsSayisiHa

\u{15c3}: ''

CanadianSyllabicsCarrierGhu

\u{15c4}: ''

CanadianSyllabicsCarrierGho

\u{15c5}: ''

CanadianSyllabicsCarrierGhe

\u{15c6}: ''

CanadianSyllabicsCarrierGhee

\u{15c7}: ''

CanadianSyllabicsCarrierGhi

\u{15c8}: ''

CanadianSyllabicsCarrierGha

\u{15c9}: ''

CanadianSyllabicsCarrierRu

\u{15ca}: ''

CanadianSyllabicsCarrierRo

\u{15cb}: ''

CanadianSyllabicsCarrierRe

\u{15cc}: ''

CanadianSyllabicsCarrierRee

\u{15cd}: ''

CanadianSyllabicsCarrierRi

\u{15ce}: ''

CanadianSyllabicsCarrierRa

\u{15cf}: ''

CanadianSyllabicsCarrierWu

\u{15d0}: ''

CanadianSyllabicsCarrierWo

\u{15d1}: ''

CanadianSyllabicsCarrierWe

\u{15d2}: ''

CanadianSyllabicsCarrierWee

\u{15d3}: ''

CanadianSyllabicsCarrierWi

\u{15d4}: ''

CanadianSyllabicsCarrierWa

\u{15d5}: ''

CanadianSyllabicsCarrierHwu

\u{15d6}: ''

CanadianSyllabicsCarrierHwo

\u{15d7}: ''

CanadianSyllabicsCarrierHwe

\u{15d8}: ''

CanadianSyllabicsCarrierHwee

\u{15d9}: ''

CanadianSyllabicsCarrierHwi

\u{15da}: ''

CanadianSyllabicsCarrierHwa

\u{15db}: ''

CanadianSyllabicsCarrierThu

\u{15dc}: ''

CanadianSyllabicsCarrierTho

\u{15dd}: ''

CanadianSyllabicsCarrierThe

\u{15de}: ''

CanadianSyllabicsCarrierThee

\u{15df}: ''

CanadianSyllabicsCarrierThi

\u{15e0}: ''

CanadianSyllabicsCarrierTha

\u{15e1}: ''

CanadianSyllabicsCarrierTtu

\u{15e2}: ''

CanadianSyllabicsCarrierTto

\u{15e3}: ''

CanadianSyllabicsCarrierTte

\u{15e4}: ''

CanadianSyllabicsCarrierTtee

\u{15e5}: ''

CanadianSyllabicsCarrierTti

\u{15e6}: ''

CanadianSyllabicsCarrierTta

\u{15e7}: ''

CanadianSyllabicsCarrierPu

\u{15e8}: ''

CanadianSyllabicsCarrierPo

\u{15e9}: ''

CanadianSyllabicsCarrierPe

\u{15ea}: ''

CanadianSyllabicsCarrierPee

\u{15eb}: ''

CanadianSyllabicsCarrierPi

\u{15ec}: ''

CanadianSyllabicsCarrierPa

\u{15ed}: ''

CanadianSyllabicsCarrierP

\u{15ee}: ''

CanadianSyllabicsCarrierGu

\u{15ef}: ''

CanadianSyllabicsCarrierGo

\u{15f0}: ''

CanadianSyllabicsCarrierGe

\u{15f1}: ''

CanadianSyllabicsCarrierGee

\u{15f2}: ''

CanadianSyllabicsCarrierGi

\u{15f3}: ''

CanadianSyllabicsCarrierGa

\u{15f4}: ''

CanadianSyllabicsCarrierKhu

\u{15f5}: ''

CanadianSyllabicsCarrierKho

\u{15f6}: ''

CanadianSyllabicsCarrierKhe

\u{15f7}: ''

CanadianSyllabicsCarrierKhee

\u{15f8}: ''

CanadianSyllabicsCarrierKhi

\u{15f9}: ''

CanadianSyllabicsCarrierKha

\u{15fa}: ''

CanadianSyllabicsCarrierKku

\u{15fb}: ''

CanadianSyllabicsCarrierKko

\u{15fc}: ''

CanadianSyllabicsCarrierKke

\u{15fd}: ''

CanadianSyllabicsCarrierKkee

\u{15fe}: ''

CanadianSyllabicsCarrierKki

\u{15ff}: ''

CanadianSyllabicsCarrierKka

\u{1600}: ''

CanadianSyllabicsCarrierKk

\u{1601}: ''

CanadianSyllabicsCarrierNu

\u{1602}: ''

CanadianSyllabicsCarrierNo

\u{1603}: ''

CanadianSyllabicsCarrierNe

\u{1604}: ''

CanadianSyllabicsCarrierNee

\u{1605}: ''

CanadianSyllabicsCarrierNi

\u{1606}: ''

CanadianSyllabicsCarrierNa

\u{1607}: ''

CanadianSyllabicsCarrierMu

\u{1608}: ''

CanadianSyllabicsCarrierMo

\u{1609}: ''

CanadianSyllabicsCarrierMe

\u{160a}: ''

CanadianSyllabicsCarrierMee

\u{160b}: ''

CanadianSyllabicsCarrierMi

\u{160c}: ''

CanadianSyllabicsCarrierMa

\u{160d}: ''

CanadianSyllabicsCarrierYu

\u{160e}: ''

CanadianSyllabicsCarrierYo

\u{160f}: ''

CanadianSyllabicsCarrierYe

\u{1610}: ''

CanadianSyllabicsCarrierYee

\u{1611}: ''

CanadianSyllabicsCarrierYi

\u{1612}: ''

CanadianSyllabicsCarrierYa

\u{1613}: ''

CanadianSyllabicsCarrierJu

\u{1614}: ''

CanadianSyllabicsSayisiJu

\u{1615}: ''

CanadianSyllabicsCarrierJo

\u{1616}: ''

CanadianSyllabicsCarrierJe

\u{1617}: ''

CanadianSyllabicsCarrierJee

\u{1618}: ''

CanadianSyllabicsCarrierJi

\u{1619}: ''

CanadianSyllabicsSayisiJi

\u{161a}: ''

CanadianSyllabicsCarrierJa

\u{161b}: ''

CanadianSyllabicsCarrierJju

\u{161c}: ''

CanadianSyllabicsCarrierJjo

\u{161d}: ''

CanadianSyllabicsCarrierJje

\u{161e}: ''

CanadianSyllabicsCarrierJjee

\u{161f}: ''

CanadianSyllabicsCarrierJji

\u{1620}: ''

CanadianSyllabicsCarrierJja

\u{1621}: ''

CanadianSyllabicsCarrierLu

\u{1622}: ''

CanadianSyllabicsCarrierLo

\u{1623}: ''

CanadianSyllabicsCarrierLe

\u{1624}: ''

CanadianSyllabicsCarrierLee

\u{1625}: ''

CanadianSyllabicsCarrierLi

\u{1626}: ''

CanadianSyllabicsCarrierLa

\u{1627}: ''

CanadianSyllabicsCarrierDlu

\u{1628}: ''

CanadianSyllabicsCarrierDlo

\u{1629}: ''

CanadianSyllabicsCarrierDle

\u{162a}: ''

CanadianSyllabicsCarrierDlee

\u{162b}: ''

CanadianSyllabicsCarrierDli

\u{162c}: ''

CanadianSyllabicsCarrierDla

\u{162d}: ''

CanadianSyllabicsCarrierLhu

\u{162e}: ''

CanadianSyllabicsCarrierLho

\u{162f}: ''

CanadianSyllabicsCarrierLhe

\u{1630}: ''

CanadianSyllabicsCarrierLhee

\u{1631}: ''

CanadianSyllabicsCarrierLhi

\u{1632}: ''

CanadianSyllabicsCarrierLha

\u{1633}: ''

CanadianSyllabicsCarrierTlhu

\u{1634}: ''

CanadianSyllabicsCarrierTlho

\u{1635}: ''

CanadianSyllabicsCarrierTlhe

\u{1636}: ''

CanadianSyllabicsCarrierTlhee

\u{1637}: ''

CanadianSyllabicsCarrierTlhi

\u{1638}: ''

CanadianSyllabicsCarrierTlha

\u{1639}: ''

CanadianSyllabicsCarrierTlu

\u{163a}: ''

CanadianSyllabicsCarrierTlo

\u{163b}: ''

CanadianSyllabicsCarrierTle

\u{163c}: ''

CanadianSyllabicsCarrierTlee

\u{163d}: ''

CanadianSyllabicsCarrierTli

\u{163e}: ''

CanadianSyllabicsCarrierTla

\u{163f}: ''

CanadianSyllabicsCarrierZu

\u{1640}: ''

CanadianSyllabicsCarrierZo

\u{1641}: ''

CanadianSyllabicsCarrierZe

\u{1642}: ''

CanadianSyllabicsCarrierZee

\u{1643}: ''

CanadianSyllabicsCarrierZi

\u{1644}: ''

CanadianSyllabicsCarrierZa

\u{1645}: ''

CanadianSyllabicsCarrierZ

\u{1646}: ''

CanadianSyllabicsCarrierInitialZ

\u{1647}: ''

CanadianSyllabicsCarrierDzu

\u{1648}: ''

CanadianSyllabicsCarrierDzo

\u{1649}: ''

CanadianSyllabicsCarrierDze

\u{164a}: ''

CanadianSyllabicsCarrierDzee

\u{164b}: ''

CanadianSyllabicsCarrierDzi

\u{164c}: ''

CanadianSyllabicsCarrierDza

\u{164d}: ''

CanadianSyllabicsCarrierSu

\u{164e}: ''

CanadianSyllabicsCarrierSo

\u{164f}: ''

CanadianSyllabicsCarrierSe

\u{1650}: ''

CanadianSyllabicsCarrierSee

\u{1651}: ''

CanadianSyllabicsCarrierSi

\u{1652}: ''

CanadianSyllabicsCarrierSa

\u{1653}: ''

CanadianSyllabicsCarrierShu

\u{1654}: ''

CanadianSyllabicsCarrierSho

\u{1655}: ''

CanadianSyllabicsCarrierShe

\u{1656}: ''

CanadianSyllabicsCarrierShee

\u{1657}: ''

CanadianSyllabicsCarrierShi

\u{1658}: ''

CanadianSyllabicsCarrierSha

\u{1659}: ''

CanadianSyllabicsCarrierSh

\u{165a}: ''

CanadianSyllabicsCarrierTsu

\u{165b}: ''

CanadianSyllabicsCarrierTso

\u{165c}: ''

CanadianSyllabicsCarrierTse

\u{165d}: ''

CanadianSyllabicsCarrierTsee

\u{165e}: ''

CanadianSyllabicsCarrierTsi

\u{165f}: ''

CanadianSyllabicsCarrierTsa

\u{1660}: ''

CanadianSyllabicsCarrierChu

\u{1661}: ''

CanadianSyllabicsCarrierCho

\u{1662}: ''

CanadianSyllabicsCarrierChe

\u{1663}: ''

CanadianSyllabicsCarrierChee

\u{1664}: ''

CanadianSyllabicsCarrierChi

\u{1665}: ''

CanadianSyllabicsCarrierCha

\u{1666}: ''

CanadianSyllabicsCarrierTtsu

\u{1667}: ''

CanadianSyllabicsCarrierTtso

\u{1668}: ''

CanadianSyllabicsCarrierTtse

\u{1669}: ''

CanadianSyllabicsCarrierTtsee

\u{166a}: ''

CanadianSyllabicsCarrierTtsi

\u{166b}: ''

CanadianSyllabicsCarrierTtsa

\u{166c}: ''

CanadianSyllabicsChiSign

\u{166d}: ''

CanadianSyllabicsFullStop

\u{166e}: ''

CanadianSyllabicsQai

\u{166f}: ''

CanadianSyllabicsNgai

\u{1670}: ''

CanadianSyllabicsNngi

\u{1671}: ''

CanadianSyllabicsNngii

\u{1672}: ''

CanadianSyllabicsNngo

\u{1673}: ''

CanadianSyllabicsNngoo

\u{1674}: ''

CanadianSyllabicsNnga

\u{1675}: ''

CanadianSyllabicsNngaa

\u{1676}: ''

CanadianSyllabicsWoodsDashCreeThwee

\u{1677}: ''

CanadianSyllabicsWoodsDashCreeThwi

\u{1678}: ''

CanadianSyllabicsWoodsDashCreeThwii

\u{1679}: ''

CanadianSyllabicsWoodsDashCreeThwo

\u{167a}: ''

CanadianSyllabicsWoodsDashCreeThwoo

\u{167b}: ''

CanadianSyllabicsWoodsDashCreeThwa

\u{167c}: ''

CanadianSyllabicsWoodsDashCreeThwaa

\u{167d}: ''

CanadianSyllabicsWoodsDashCreeFinalTh

\u{167e}: ''

Methods

impl UnifiedCanadianAboriginalSyllabics

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case