Enum unicode_types::vai::Vai

pub enum Vai {
    SyllableEe,
    SyllableEen,
    SyllableHee,
    SyllableWee,
    SyllableWeen,
    SyllablePee,
    SyllableBhee,
    SyllableBee,
    SyllableMbee,
    SyllableKpee,
    SyllableMgbee,
    SyllableGbee,
    SyllableFee,
    SyllableVee,
    SyllableTee,
    SyllableThee,
    SyllableDhee,
    SyllableDhhee,
    SyllableLee,
    SyllableRee,
    SyllableDee,
    SyllableNdee,
    SyllableSee,
    SyllableShee,
    SyllableZee,
    SyllableZhee,
    SyllableCee,
    SyllableJee,
    SyllableNjee,
    SyllableYee,
    SyllableKee,
    SyllableNggee,
    SyllableGee,
    SyllableMee,
    SyllableNee,
    SyllableNyee,
    SyllableI,
    SyllableIn,
    SyllableHi,
    SyllableHin,
    SyllableWi,
    SyllableWin,
    SyllablePi,
    SyllableBhi,
    SyllableBi,
    SyllableMbi,
    SyllableKpi,
    SyllableMgbi,
    SyllableGbi,
    SyllableFi,
    SyllableVi,
    SyllableTi,
    SyllableThi,
    SyllableDhi,
    SyllableDhhi,
    SyllableLi,
    SyllableRi,
    SyllableDi,
    SyllableNdi,
    SyllableSi,
    SyllableShi,
    SyllableZi,
    SyllableZhi,
    SyllableCi,
    SyllableJi,
    SyllableNji,
    SyllableYi,
    SyllableKi,
    SyllableNggi,
    SyllableGi,
    SyllableMi,
    SyllableNi,
    SyllableNyi,
    SyllableA,
    SyllableAn,
    SyllableNgan,
    SyllableHa,
    SyllableHan,
    SyllableWa,
    SyllableWan,
    SyllablePa,
    SyllableBha,
    SyllableBa,
    SyllableMba,
    SyllableKpa,
    SyllableKpan,
    SyllableMgba,
    SyllableGba,
    SyllableFa,
    SyllableVa,
    SyllableTa,
    SyllableTha,
    SyllableDha,
    SyllableDhha,
    SyllableLa,
    SyllableRa,
    SyllableDa,
    SyllableNda,
    SyllableSa,
    SyllableSha,
    SyllableZa,
    SyllableZha,
    SyllableCa,
    SyllableJa,
    SyllableNja,
    SyllableYa,
    SyllableKa,
    SyllableKan,
    SyllableNgga,
    SyllableGa,
    SyllableMa,
    SyllableNa,
    SyllableNya,
    SyllableOo,
    SyllableOon,
    SyllableHoo,
    SyllableWoo,
    SyllableWoon,
    SyllablePoo,
    SyllableBhoo,
    SyllableBoo,
    SyllableMboo,
    SyllableKpoo,
    SyllableMgboo,
    SyllableGboo,
    SyllableFoo,
    SyllableVoo,
    SyllableToo,
    SyllableThoo,
    SyllableDhoo,
    SyllableDhhoo,
    SyllableLoo,
    SyllableRoo,
    SyllableDoo,
    SyllableNdoo,
    SyllableSoo,
    SyllableShoo,
    SyllableZoo,
    SyllableZhoo,
    SyllableCoo,
    SyllableJoo,
    SyllableNjoo,
    SyllableYoo,
    SyllableKoo,
    SyllableNggoo,
    SyllableGoo,
    SyllableMoo,
    SyllableNoo,
    SyllableNyoo,
    SyllableU,
    SyllableUn,
    SyllableHu,
    SyllableHun,
    SyllableWu,
    SyllableWun,
    SyllablePu,
    SyllableBhu,
    SyllableBu,
    SyllableMbu,
    SyllableKpu,
    SyllableMgbu,
    SyllableGbu,
    SyllableFu,
    SyllableVu,
    SyllableTu,
    SyllableThu,
    SyllableDhu,
    SyllableDhhu,
    SyllableLu,
    SyllableRu,
    SyllableDu,
    SyllableNdu,
    SyllableSu,
    SyllableShu,
    SyllableZu,
    SyllableZhu,
    SyllableCu,
    SyllableJu,
    SyllableNju,
    SyllableYu,
    SyllableKu,
    SyllableNggu,
    SyllableGu,
    SyllableMu,
    SyllableNu,
    SyllableNyu,
    SyllableO,
    SyllableOn,
    SyllableNgon,
    SyllableHo,
    SyllableHon,
    SyllableWo,
    SyllableWon,
    SyllablePo,
    SyllableBho,
    SyllableBo,
    SyllableMbo,
    SyllableKpo,
    SyllableMgbo,
    SyllableGbo,
    SyllableGbon,
    SyllableFo,
    SyllableVo,
    SyllableTo,
    SyllableTho,
    SyllableDho,
    SyllableDhho,
    SyllableLo,
    SyllableRo,
    SyllableDo,
    SyllableNdo,
    SyllableSo,
    SyllableSho,
    SyllableZo,
    SyllableZho,
    SyllableCo,
    SyllableJo,
    SyllableNjo,
    SyllableYo,
    SyllableKo,
    SyllableNggo,
    SyllableGo,
    SyllableMo,
    SyllableNo,
    SyllableNyo,
    SyllableE,
    SyllableEn,
    SyllableNgen,
    SyllableHe,
    SyllableHen,
    SyllableWe,
    SyllableWen,
    SyllablePe,
    SyllableBhe,
    SyllableBe,
    SyllableMbe,
    SyllableKpe,
    SyllableKpen,
    SyllableMgbe,
    SyllableGbe,
    SyllableGben,
    SyllableFe,
    SyllableVe,
    SyllableTe,
    SyllableThe,
    SyllableDhe,
    SyllableDhhe,
    SyllableLe,
    SyllableRe,
    SyllableDe,
    SyllableNde,
    SyllableSe,
    SyllableShe,
    SyllableZe,
    SyllableZhe,
    SyllableCe,
    SyllableJe,
    SyllableNje,
    SyllableYe,
    SyllableKe,
    SyllableNgge,
    SyllableNggen,
    SyllableGe,
    SyllableGen,
    SyllableMe,
    SyllableNe,
    SyllableNye,
    SyllableNg,
    SyllableLengthener,
    Comma,
    FullStop,
    QuestionMark,
    SyllableNdoleFa,
    SyllableNdoleKa,
    SyllableNdoleSoo,
    SymbolFeeng,
    SymbolKeeng,
    SymbolTing,
    SymbolNii,
    SymbolBang,
    SymbolFaa,
    SymbolTaa,
    SymbolDang,
    SymbolDoong,
    SymbolKung,
    SymbolTong,
    SymbolDoDashO,
    SymbolJong,
    DigitZero,
    DigitOne,
    DigitTwo,
    DigitThree,
    DigitFour,
    DigitFive,
    DigitSix,
    DigitSeven,
    DigitEight,
    DigitNine,
    SyllableNdoleMa,
    SyllableNdoleDo,
}

An enum to represent all characters in the Vai block.

Variants
SyllableEe

\u{a500}: 'ꔀ'

SyllableEen

\u{a501}: 'ꔁ'

SyllableHee

\u{a502}: 'ꔂ'

SyllableWee

\u{a503}: 'ꔃ'

SyllableWeen

\u{a504}: 'ꔄ'

SyllablePee

\u{a505}: 'ꔅ'

SyllableBhee

\u{a506}: 'ꔆ'

SyllableBee

\u{a507}: 'ꔇ'

SyllableMbee

\u{a508}: 'ꔈ'

SyllableKpee

\u{a509}: 'ꔉ'

SyllableMgbee

\u{a50a}: 'ꔊ'

SyllableGbee

\u{a50b}: 'ꔋ'

SyllableFee

\u{a50c}: 'ꔌ'

SyllableVee

\u{a50d}: 'ꔍ'

SyllableTee

\u{a50e}: 'ꔎ'

SyllableThee

\u{a50f}: 'ꔏ'

SyllableDhee

\u{a510}: 'ꔐ'

SyllableDhhee

\u{a511}: 'ꔑ'

SyllableLee

\u{a512}: 'ꔒ'

SyllableRee

\u{a513}: 'ꔓ'

SyllableDee

\u{a514}: 'ꔔ'

SyllableNdee

\u{a515}: 'ꔕ'

SyllableSee

\u{a516}: 'ꔖ'

SyllableShee

\u{a517}: 'ꔗ'

SyllableZee

\u{a518}: 'ꔘ'

SyllableZhee

\u{a519}: 'ꔙ'

SyllableCee

\u{a51a}: 'ꔚ'

SyllableJee

\u{a51b}: 'ꔛ'

SyllableNjee

\u{a51c}: 'ꔜ'

SyllableYee

\u{a51d}: 'ꔝ'

SyllableKee

\u{a51e}: 'ꔞ'

SyllableNggee

\u{a51f}: 'ꔟ'

SyllableGee

\u{a520}: 'ꔠ'

SyllableMee

\u{a521}: 'ꔡ'

SyllableNee

\u{a522}: 'ꔢ'

SyllableNyee

\u{a523}: 'ꔣ'

SyllableI

\u{a524}: 'ꔤ'

SyllableIn

\u{a525}: 'ꔥ'

SyllableHi

\u{a526}: 'ꔦ'

SyllableHin

\u{a527}: 'ꔧ'

SyllableWi

\u{a528}: 'ꔨ'

SyllableWin

\u{a529}: 'ꔩ'

SyllablePi

\u{a52a}: 'ꔪ'

SyllableBhi

\u{a52b}: 'ꔫ'

SyllableBi

\u{a52c}: 'ꔬ'

SyllableMbi

\u{a52d}: 'ꔭ'

SyllableKpi

\u{a52e}: 'ꔮ'

SyllableMgbi

\u{a52f}: 'ꔯ'

SyllableGbi

\u{a530}: 'ꔰ'

SyllableFi

\u{a531}: 'ꔱ'

SyllableVi

\u{a532}: 'ꔲ'

SyllableTi

\u{a533}: 'ꔳ'

SyllableThi

\u{a534}: 'ꔴ'

SyllableDhi

\u{a535}: 'ꔵ'

SyllableDhhi

\u{a536}: 'ꔶ'

SyllableLi

\u{a537}: 'ꔷ'

SyllableRi

\u{a538}: 'ꔸ'

SyllableDi

\u{a539}: 'ꔹ'

SyllableNdi

\u{a53a}: 'ꔺ'

SyllableSi

\u{a53b}: 'ꔻ'

SyllableShi

\u{a53c}: 'ꔼ'

SyllableZi

\u{a53d}: 'ꔽ'

SyllableZhi

\u{a53e}: 'ꔾ'

SyllableCi

\u{a53f}: 'ꔿ'

SyllableJi

\u{a540}: 'ꕀ'

SyllableNji

\u{a541}: 'ꕁ'

SyllableYi

\u{a542}: 'ꕂ'

SyllableKi

\u{a543}: 'ꕃ'

SyllableNggi

\u{a544}: 'ꕄ'

SyllableGi

\u{a545}: 'ꕅ'

SyllableMi

\u{a546}: 'ꕆ'

SyllableNi

\u{a547}: 'ꕇ'

SyllableNyi

\u{a548}: 'ꕈ'

SyllableA

\u{a549}: 'ꕉ'

SyllableAn

\u{a54a}: 'ꕊ'

SyllableNgan

\u{a54b}: 'ꕋ'

SyllableHa

\u{a54c}: 'ꕌ'

SyllableHan

\u{a54d}: 'ꕍ'

SyllableWa

\u{a54e}: 'ꕎ'

SyllableWan

\u{a54f}: 'ꕏ'

SyllablePa

\u{a550}: 'ꕐ'

SyllableBha

\u{a551}: 'ꕑ'

SyllableBa

\u{a552}: 'ꕒ'

SyllableMba

\u{a553}: 'ꕓ'

SyllableKpa

\u{a554}: 'ꕔ'

SyllableKpan

\u{a555}: 'ꕕ'

SyllableMgba

\u{a556}: 'ꕖ'

SyllableGba

\u{a557}: 'ꕗ'

SyllableFa

\u{a558}: 'ꕘ'

SyllableVa

\u{a559}: 'ꕙ'

SyllableTa

\u{a55a}: 'ꕚ'

SyllableTha

\u{a55b}: 'ꕛ'

SyllableDha

\u{a55c}: 'ꕜ'

SyllableDhha

\u{a55d}: 'ꕝ'

SyllableLa

\u{a55e}: 'ꕞ'

SyllableRa

\u{a55f}: 'ꕟ'

SyllableDa

\u{a560}: 'ꕠ'

SyllableNda

\u{a561}: 'ꕡ'

SyllableSa

\u{a562}: 'ꕢ'

SyllableSha

\u{a563}: 'ꕣ'

SyllableZa

\u{a564}: 'ꕤ'

SyllableZha

\u{a565}: 'ꕥ'

SyllableCa

\u{a566}: 'ꕦ'

SyllableJa

\u{a567}: 'ꕧ'

SyllableNja

\u{a568}: 'ꕨ'

SyllableYa

\u{a569}: 'ꕩ'

SyllableKa

\u{a56a}: 'ꕪ'

SyllableKan

\u{a56b}: 'ꕫ'

SyllableNgga

\u{a56c}: 'ꕬ'

SyllableGa

\u{a56d}: 'ꕭ'

SyllableMa

\u{a56e}: 'ꕮ'

SyllableNa

\u{a56f}: 'ꕯ'

SyllableNya

\u{a570}: 'ꕰ'

SyllableOo

\u{a571}: 'ꕱ'

SyllableOon

\u{a572}: 'ꕲ'

SyllableHoo

\u{a573}: 'ꕳ'

SyllableWoo

\u{a574}: 'ꕴ'

SyllableWoon

\u{a575}: 'ꕵ'

SyllablePoo

\u{a576}: 'ꕶ'

SyllableBhoo

\u{a577}: 'ꕷ'

SyllableBoo

\u{a578}: 'ꕸ'

SyllableMboo

\u{a579}: 'ꕹ'

SyllableKpoo

\u{a57a}: 'ꕺ'

SyllableMgboo

\u{a57b}: 'ꕻ'

SyllableGboo

\u{a57c}: 'ꕼ'

SyllableFoo

\u{a57d}: 'ꕽ'

SyllableVoo

\u{a57e}: 'ꕾ'

SyllableToo

\u{a57f}: 'ꕿ'

SyllableThoo

\u{a580}: 'ꖀ'

SyllableDhoo

\u{a581}: 'ꖁ'

SyllableDhhoo

\u{a582}: 'ꖂ'

SyllableLoo

\u{a583}: 'ꖃ'

SyllableRoo

\u{a584}: 'ꖄ'

SyllableDoo

\u{a585}: 'ꖅ'

SyllableNdoo

\u{a586}: 'ꖆ'

SyllableSoo

\u{a587}: 'ꖇ'

SyllableShoo

\u{a588}: 'ꖈ'

SyllableZoo

\u{a589}: 'ꖉ'

SyllableZhoo

\u{a58a}: 'ꖊ'

SyllableCoo

\u{a58b}: 'ꖋ'

SyllableJoo

\u{a58c}: 'ꖌ'

SyllableNjoo

\u{a58d}: 'ꖍ'

SyllableYoo

\u{a58e}: 'ꖎ'

SyllableKoo

\u{a58f}: 'ꖏ'

SyllableNggoo

\u{a590}: 'ꖐ'

SyllableGoo

\u{a591}: 'ꖑ'

SyllableMoo

\u{a592}: 'ꖒ'

SyllableNoo

\u{a593}: 'ꖓ'

SyllableNyoo

\u{a594}: 'ꖔ'

SyllableU

\u{a595}: 'ꖕ'

SyllableUn

\u{a596}: 'ꖖ'

SyllableHu

\u{a597}: 'ꖗ'

SyllableHun

\u{a598}: 'ꖘ'

SyllableWu

\u{a599}: 'ꖙ'

SyllableWun

\u{a59a}: 'ꖚ'

SyllablePu

\u{a59b}: 'ꖛ'

SyllableBhu

\u{a59c}: 'ꖜ'

SyllableBu

\u{a59d}: 'ꖝ'

SyllableMbu

\u{a59e}: 'ꖞ'

SyllableKpu

\u{a59f}: 'ꖟ'

SyllableMgbu

\u{a5a0}: 'ꖠ'

SyllableGbu

\u{a5a1}: 'ꖡ'

SyllableFu

\u{a5a2}: 'ꖢ'

SyllableVu

\u{a5a3}: 'ꖣ'

SyllableTu

\u{a5a4}: 'ꖤ'

SyllableThu

\u{a5a5}: 'ꖥ'

SyllableDhu

\u{a5a6}: 'ꖦ'

SyllableDhhu

\u{a5a7}: 'ꖧ'

SyllableLu

\u{a5a8}: 'ꖨ'

SyllableRu

\u{a5a9}: 'ꖩ'

SyllableDu

\u{a5aa}: 'ꖪ'

SyllableNdu

\u{a5ab}: 'ꖫ'

SyllableSu

\u{a5ac}: 'ꖬ'

SyllableShu

\u{a5ad}: 'ꖭ'

SyllableZu

\u{a5ae}: 'ꖮ'

SyllableZhu

\u{a5af}: 'ꖯ'

SyllableCu

\u{a5b0}: 'ꖰ'

SyllableJu

\u{a5b1}: 'ꖱ'

SyllableNju

\u{a5b2}: 'ꖲ'

SyllableYu

\u{a5b3}: 'ꖳ'

SyllableKu

\u{a5b4}: 'ꖴ'

SyllableNggu

\u{a5b5}: 'ꖵ'

SyllableGu

\u{a5b6}: 'ꖶ'

SyllableMu

\u{a5b7}: 'ꖷ'

SyllableNu

\u{a5b8}: 'ꖸ'

SyllableNyu

\u{a5b9}: 'ꖹ'

SyllableO

\u{a5ba}: 'ꖺ'

SyllableOn

\u{a5bb}: 'ꖻ'

SyllableNgon

\u{a5bc}: 'ꖼ'

SyllableHo

\u{a5bd}: 'ꖽ'

SyllableHon

\u{a5be}: 'ꖾ'

SyllableWo

\u{a5bf}: 'ꖿ'

SyllableWon

\u{a5c0}: 'ꗀ'

SyllablePo

\u{a5c1}: 'ꗁ'

SyllableBho

\u{a5c2}: 'ꗂ'

SyllableBo

\u{a5c3}: 'ꗃ'

SyllableMbo

\u{a5c4}: 'ꗄ'

SyllableKpo

\u{a5c5}: 'ꗅ'

SyllableMgbo

\u{a5c6}: 'ꗆ'

SyllableGbo

\u{a5c7}: 'ꗇ'

SyllableGbon

\u{a5c8}: 'ꗈ'

SyllableFo

\u{a5c9}: 'ꗉ'

SyllableVo

\u{a5ca}: 'ꗊ'

SyllableTo

\u{a5cb}: 'ꗋ'

SyllableTho

\u{a5cc}: 'ꗌ'

SyllableDho

\u{a5cd}: 'ꗍ'

SyllableDhho

\u{a5ce}: 'ꗎ'

SyllableLo

\u{a5cf}: 'ꗏ'

SyllableRo

\u{a5d0}: 'ꗐ'

SyllableDo

\u{a5d1}: 'ꗑ'

SyllableNdo

\u{a5d2}: 'ꗒ'

SyllableSo

\u{a5d3}: 'ꗓ'

SyllableSho

\u{a5d4}: 'ꗔ'

SyllableZo

\u{a5d5}: 'ꗕ'

SyllableZho

\u{a5d6}: 'ꗖ'

SyllableCo

\u{a5d7}: 'ꗗ'

SyllableJo

\u{a5d8}: 'ꗘ'

SyllableNjo

\u{a5d9}: 'ꗙ'

SyllableYo

\u{a5da}: 'ꗚ'

SyllableKo

\u{a5db}: 'ꗛ'

SyllableNggo

\u{a5dc}: 'ꗜ'

SyllableGo

\u{a5dd}: 'ꗝ'

SyllableMo

\u{a5de}: 'ꗞ'

SyllableNo

\u{a5df}: 'ꗟ'

SyllableNyo

\u{a5e0}: 'ꗠ'

SyllableE

\u{a5e1}: 'ꗡ'

SyllableEn

\u{a5e2}: 'ꗢ'

SyllableNgen

\u{a5e3}: 'ꗣ'

SyllableHe

\u{a5e4}: 'ꗤ'

SyllableHen

\u{a5e5}: 'ꗥ'

SyllableWe

\u{a5e6}: 'ꗦ'

SyllableWen

\u{a5e7}: 'ꗧ'

SyllablePe

\u{a5e8}: 'ꗨ'

SyllableBhe

\u{a5e9}: 'ꗩ'

SyllableBe

\u{a5ea}: 'ꗪ'

SyllableMbe

\u{a5eb}: 'ꗫ'

SyllableKpe

\u{a5ec}: 'ꗬ'

SyllableKpen

\u{a5ed}: 'ꗭ'

SyllableMgbe

\u{a5ee}: 'ꗮ'

SyllableGbe

\u{a5ef}: 'ꗯ'

SyllableGben

\u{a5f0}: 'ꗰ'

SyllableFe

\u{a5f1}: 'ꗱ'

SyllableVe

\u{a5f2}: 'ꗲ'

SyllableTe

\u{a5f3}: 'ꗳ'

SyllableThe

\u{a5f4}: 'ꗴ'

SyllableDhe

\u{a5f5}: 'ꗵ'

SyllableDhhe

\u{a5f6}: 'ꗶ'

SyllableLe

\u{a5f7}: 'ꗷ'

SyllableRe

\u{a5f8}: 'ꗸ'

SyllableDe

\u{a5f9}: 'ꗹ'

SyllableNde

\u{a5fa}: 'ꗺ'

SyllableSe

\u{a5fb}: 'ꗻ'

SyllableShe

\u{a5fc}: 'ꗼ'

SyllableZe

\u{a5fd}: 'ꗽ'

SyllableZhe

\u{a5fe}: 'ꗾ'

SyllableCe

\u{a5ff}: 'ꗿ'

SyllableJe

\u{a600}: 'ꘀ'

SyllableNje

\u{a601}: 'ꘁ'

SyllableYe

\u{a602}: 'ꘂ'

SyllableKe

\u{a603}: 'ꘃ'

SyllableNgge

\u{a604}: 'ꘄ'

SyllableNggen

\u{a605}: 'ꘅ'

SyllableGe

\u{a606}: 'ꘆ'

SyllableGen

\u{a607}: 'ꘇ'

SyllableMe

\u{a608}: 'ꘈ'

SyllableNe

\u{a609}: 'ꘉ'

SyllableNye

\u{a60a}: 'ꘊ'

SyllableNg

\u{a60b}: 'ꘋ'

SyllableLengthener

\u{a60c}: 'ꘌ'

Comma

\u{a60d}: '꘍'

FullStop

\u{a60e}: '꘎'

QuestionMark

\u{a60f}: '꘏'

SyllableNdoleFa

\u{a610}: 'ꘐ'

SyllableNdoleKa

\u{a611}: 'ꘑ'

SyllableNdoleSoo

\u{a612}: 'ꘒ'

SymbolFeeng

\u{a613}: 'ꘓ'

SymbolKeeng

\u{a614}: 'ꘔ'

SymbolTing

\u{a615}: 'ꘕ'

SymbolNii

\u{a616}: 'ꘖ'

SymbolBang

\u{a617}: 'ꘗ'

SymbolFaa

\u{a618}: 'ꘘ'

SymbolTaa

\u{a619}: 'ꘙ'

SymbolDang

\u{a61a}: 'ꘚ'

SymbolDoong

\u{a61b}: 'ꘛ'

SymbolKung

\u{a61c}: 'ꘜ'

SymbolTong

\u{a61d}: 'ꘝ'

SymbolDoDashO

\u{a61e}: 'ꘞ'

SymbolJong

\u{a61f}: 'ꘟ'

DigitZero

\u{a620}: '꘠'

DigitOne

\u{a621}: '꘡'

DigitTwo

\u{a622}: '꘢'

DigitThree

\u{a623}: '꘣'

DigitFour

\u{a624}: '꘤'

DigitFive

\u{a625}: '꘥'

DigitSix

\u{a626}: '꘦'

DigitSeven

\u{a627}: '꘧'

DigitEight

\u{a628}: '꘨'

DigitNine

\u{a629}: '꘩'

SyllableNdoleMa

\u{a62a}: 'ꘪ'

SyllableNdoleDo

\u{a62b}: 'ꘫ'

Methods

impl Vai

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case