Enum unicode_types::variation_selectors_supplement::VariationSelectorsSupplement

pub enum VariationSelectorsSupplement {
    VariationSelectorDash17,
    VariationSelectorDash18,
    VariationSelectorDash19,
    VariationSelectorDash20,
    VariationSelectorDash21,
    VariationSelectorDash22,
    VariationSelectorDash23,
    VariationSelectorDash24,
    VariationSelectorDash25,
    VariationSelectorDash26,
    VariationSelectorDash27,
    VariationSelectorDash28,
    VariationSelectorDash29,
    VariationSelectorDash30,
    VariationSelectorDash31,
    VariationSelectorDash32,
    VariationSelectorDash33,
    VariationSelectorDash34,
    VariationSelectorDash35,
    VariationSelectorDash36,
    VariationSelectorDash37,
    VariationSelectorDash38,
    VariationSelectorDash39,
    VariationSelectorDash40,
    VariationSelectorDash41,
    VariationSelectorDash42,
    VariationSelectorDash43,
    VariationSelectorDash44,
    VariationSelectorDash45,
    VariationSelectorDash46,
    VariationSelectorDash47,
    VariationSelectorDash48,
    VariationSelectorDash49,
    VariationSelectorDash50,
    VariationSelectorDash51,
    VariationSelectorDash52,
    VariationSelectorDash53,
    VariationSelectorDash54,
    VariationSelectorDash55,
    VariationSelectorDash56,
    VariationSelectorDash57,
    VariationSelectorDash58,
    VariationSelectorDash59,
    VariationSelectorDash60,
    VariationSelectorDash61,
    VariationSelectorDash62,
    VariationSelectorDash63,
    VariationSelectorDash64,
    VariationSelectorDash65,
    VariationSelectorDash66,
    VariationSelectorDash67,
    VariationSelectorDash68,
    VariationSelectorDash69,
    VariationSelectorDash70,
    VariationSelectorDash71,
    VariationSelectorDash72,
    VariationSelectorDash73,
    VariationSelectorDash74,
    VariationSelectorDash75,
    VariationSelectorDash76,
    VariationSelectorDash77,
    VariationSelectorDash78,
    VariationSelectorDash79,
    VariationSelectorDash80,
    VariationSelectorDash81,
    VariationSelectorDash82,
    VariationSelectorDash83,
    VariationSelectorDash84,
    VariationSelectorDash85,
    VariationSelectorDash86,
    VariationSelectorDash87,
    VariationSelectorDash88,
    VariationSelectorDash89,
    VariationSelectorDash90,
    VariationSelectorDash91,
    VariationSelectorDash92,
    VariationSelectorDash93,
    VariationSelectorDash94,
    VariationSelectorDash95,
    VariationSelectorDash96,
    VariationSelectorDash97,
    VariationSelectorDash98,
    VariationSelectorDash99,
    VariationSelectorDash100,
    VariationSelectorDash101,
    VariationSelectorDash102,
    VariationSelectorDash103,
    VariationSelectorDash104,
    VariationSelectorDash105,
    VariationSelectorDash106,
    VariationSelectorDash107,
    VariationSelectorDash108,
    VariationSelectorDash109,
    VariationSelectorDash110,
    VariationSelectorDash111,
    VariationSelectorDash112,
    VariationSelectorDash113,
    VariationSelectorDash114,
    VariationSelectorDash115,
    VariationSelectorDash116,
    VariationSelectorDash117,
    VariationSelectorDash118,
    VariationSelectorDash119,
    VariationSelectorDash120,
    VariationSelectorDash121,
    VariationSelectorDash122,
    VariationSelectorDash123,
    VariationSelectorDash124,
    VariationSelectorDash125,
    VariationSelectorDash126,
    VariationSelectorDash127,
    VariationSelectorDash128,
    VariationSelectorDash129,
    VariationSelectorDash130,
    VariationSelectorDash131,
    VariationSelectorDash132,
    VariationSelectorDash133,
    VariationSelectorDash134,
    VariationSelectorDash135,
    VariationSelectorDash136,
    VariationSelectorDash137,
    VariationSelectorDash138,
    VariationSelectorDash139,
    VariationSelectorDash140,
    VariationSelectorDash141,
    VariationSelectorDash142,
    VariationSelectorDash143,
    VariationSelectorDash144,
    VariationSelectorDash145,
    VariationSelectorDash146,
    VariationSelectorDash147,
    VariationSelectorDash148,
    VariationSelectorDash149,
    VariationSelectorDash150,
    VariationSelectorDash151,
    VariationSelectorDash152,
    VariationSelectorDash153,
    VariationSelectorDash154,
    VariationSelectorDash155,
    VariationSelectorDash156,
    VariationSelectorDash157,
    VariationSelectorDash158,
    VariationSelectorDash159,
    VariationSelectorDash160,
    VariationSelectorDash161,
    VariationSelectorDash162,
    VariationSelectorDash163,
    VariationSelectorDash164,
    VariationSelectorDash165,
    VariationSelectorDash166,
    VariationSelectorDash167,
    VariationSelectorDash168,
    VariationSelectorDash169,
    VariationSelectorDash170,
    VariationSelectorDash171,
    VariationSelectorDash172,
    VariationSelectorDash173,
    VariationSelectorDash174,
    VariationSelectorDash175,
    VariationSelectorDash176,
    VariationSelectorDash177,
    VariationSelectorDash178,
    VariationSelectorDash179,
    VariationSelectorDash180,
    VariationSelectorDash181,
    VariationSelectorDash182,
    VariationSelectorDash183,
    VariationSelectorDash184,
    VariationSelectorDash185,
    VariationSelectorDash186,
    VariationSelectorDash187,
    VariationSelectorDash188,
    VariationSelectorDash189,
    VariationSelectorDash190,
    VariationSelectorDash191,
    VariationSelectorDash192,
    VariationSelectorDash193,
    VariationSelectorDash194,
    VariationSelectorDash195,
    VariationSelectorDash196,
    VariationSelectorDash197,
    VariationSelectorDash198,
    VariationSelectorDash199,
    VariationSelectorDash200,
    VariationSelectorDash201,
    VariationSelectorDash202,
    VariationSelectorDash203,
    VariationSelectorDash204,
    VariationSelectorDash205,
    VariationSelectorDash206,
    VariationSelectorDash207,
    VariationSelectorDash208,
    VariationSelectorDash209,
    VariationSelectorDash210,
    VariationSelectorDash211,
    VariationSelectorDash212,
    VariationSelectorDash213,
    VariationSelectorDash214,
    VariationSelectorDash215,
    VariationSelectorDash216,
    VariationSelectorDash217,
    VariationSelectorDash218,
    VariationSelectorDash219,
    VariationSelectorDash220,
    VariationSelectorDash221,
    VariationSelectorDash222,
    VariationSelectorDash223,
    VariationSelectorDash224,
    VariationSelectorDash225,
    VariationSelectorDash226,
    VariationSelectorDash227,
    VariationSelectorDash228,
    VariationSelectorDash229,
    VariationSelectorDash230,
    VariationSelectorDash231,
    VariationSelectorDash232,
    VariationSelectorDash233,
    VariationSelectorDash234,
    VariationSelectorDash235,
    VariationSelectorDash236,
    VariationSelectorDash237,
    VariationSelectorDash238,
    VariationSelectorDash239,
    VariationSelectorDash240,
    VariationSelectorDash241,
    VariationSelectorDash242,
    VariationSelectorDash243,
    VariationSelectorDash244,
    VariationSelectorDash245,
    VariationSelectorDash246,
    VariationSelectorDash247,
    VariationSelectorDash248,
    VariationSelectorDash249,
    VariationSelectorDash250,
    VariationSelectorDash251,
    VariationSelectorDash252,
    VariationSelectorDash253,
    VariationSelectorDash254,
    VariationSelectorDash255,
}

An enum to represent all characters in the VariationSelectorsSupplement block.

Variants
VariationSelectorDash17

\u{e0100}: '󠄀'

VariationSelectorDash18

\u{e0101}: '󠄁'

VariationSelectorDash19

\u{e0102}: '󠄂'

VariationSelectorDash20

\u{e0103}: '󠄃'

VariationSelectorDash21

\u{e0104}: '󠄄'

VariationSelectorDash22

\u{e0105}: '󠄅'

VariationSelectorDash23

\u{e0106}: '󠄆'

VariationSelectorDash24

\u{e0107}: '󠄇'

VariationSelectorDash25

\u{e0108}: '󠄈'

VariationSelectorDash26

\u{e0109}: '󠄉'

VariationSelectorDash27

\u{e010a}: '󠄊'

VariationSelectorDash28

\u{e010b}: '󠄋'

VariationSelectorDash29

\u{e010c}: '󠄌'

VariationSelectorDash30

\u{e010d}: '󠄍'

VariationSelectorDash31

\u{e010e}: '󠄎'

VariationSelectorDash32

\u{e010f}: '󠄏'

VariationSelectorDash33

\u{e0110}: '󠄐'

VariationSelectorDash34

\u{e0111}: '󠄑'

VariationSelectorDash35

\u{e0112}: '󠄒'

VariationSelectorDash36

\u{e0113}: '󠄓'

VariationSelectorDash37

\u{e0114}: '󠄔'

VariationSelectorDash38

\u{e0115}: '󠄕'

VariationSelectorDash39

\u{e0116}: '󠄖'

VariationSelectorDash40

\u{e0117}: '󠄗'

VariationSelectorDash41

\u{e0118}: '󠄘'

VariationSelectorDash42

\u{e0119}: '󠄙'

VariationSelectorDash43

\u{e011a}: '󠄚'

VariationSelectorDash44

\u{e011b}: '󠄛'

VariationSelectorDash45

\u{e011c}: '󠄜'

VariationSelectorDash46

\u{e011d}: '󠄝'

VariationSelectorDash47

\u{e011e}: '󠄞'

VariationSelectorDash48

\u{e011f}: '󠄟'

VariationSelectorDash49

\u{e0120}: '󠄠'

VariationSelectorDash50

\u{e0121}: '󠄡'

VariationSelectorDash51

\u{e0122}: '󠄢'

VariationSelectorDash52

\u{e0123}: '󠄣'

VariationSelectorDash53

\u{e0124}: '󠄤'

VariationSelectorDash54

\u{e0125}: '󠄥'

VariationSelectorDash55

\u{e0126}: '󠄦'

VariationSelectorDash56

\u{e0127}: '󠄧'

VariationSelectorDash57

\u{e0128}: '󠄨'

VariationSelectorDash58

\u{e0129}: '󠄩'

VariationSelectorDash59

\u{e012a}: '󠄪'

VariationSelectorDash60

\u{e012b}: '󠄫'

VariationSelectorDash61

\u{e012c}: '󠄬'

VariationSelectorDash62

\u{e012d}: '󠄭'

VariationSelectorDash63

\u{e012e}: '󠄮'

VariationSelectorDash64

\u{e012f}: '󠄯'

VariationSelectorDash65

\u{e0130}: '󠄰'

VariationSelectorDash66

\u{e0131}: '󠄱'

VariationSelectorDash67

\u{e0132}: '󠄲'

VariationSelectorDash68

\u{e0133}: '󠄳'

VariationSelectorDash69

\u{e0134}: '󠄴'

VariationSelectorDash70

\u{e0135}: '󠄵'

VariationSelectorDash71

\u{e0136}: '󠄶'

VariationSelectorDash72

\u{e0137}: '󠄷'

VariationSelectorDash73

\u{e0138}: '󠄸'

VariationSelectorDash74

\u{e0139}: '󠄹'

VariationSelectorDash75

\u{e013a}: '󠄺'

VariationSelectorDash76

\u{e013b}: '󠄻'

VariationSelectorDash77

\u{e013c}: '󠄼'

VariationSelectorDash78

\u{e013d}: '󠄽'

VariationSelectorDash79

\u{e013e}: '󠄾'

VariationSelectorDash80

\u{e013f}: '󠄿'

VariationSelectorDash81

\u{e0140}: '󠅀'

VariationSelectorDash82

\u{e0141}: '󠅁'

VariationSelectorDash83

\u{e0142}: '󠅂'

VariationSelectorDash84

\u{e0143}: '󠅃'

VariationSelectorDash85

\u{e0144}: '󠅄'

VariationSelectorDash86

\u{e0145}: '󠅅'

VariationSelectorDash87

\u{e0146}: '󠅆'

VariationSelectorDash88

\u{e0147}: '󠅇'

VariationSelectorDash89

\u{e0148}: '󠅈'

VariationSelectorDash90

\u{e0149}: '󠅉'

VariationSelectorDash91

\u{e014a}: '󠅊'

VariationSelectorDash92

\u{e014b}: '󠅋'

VariationSelectorDash93

\u{e014c}: '󠅌'

VariationSelectorDash94

\u{e014d}: '󠅍'

VariationSelectorDash95

\u{e014e}: '󠅎'

VariationSelectorDash96

\u{e014f}: '󠅏'

VariationSelectorDash97

\u{e0150}: '󠅐'

VariationSelectorDash98

\u{e0151}: '󠅑'

VariationSelectorDash99

\u{e0152}: '󠅒'

VariationSelectorDash100

\u{e0153}: '󠅓'

VariationSelectorDash101

\u{e0154}: '󠅔'

VariationSelectorDash102

\u{e0155}: '󠅕'

VariationSelectorDash103

\u{e0156}: '󠅖'

VariationSelectorDash104

\u{e0157}: '󠅗'

VariationSelectorDash105

\u{e0158}: '󠅘'

VariationSelectorDash106

\u{e0159}: '󠅙'

VariationSelectorDash107

\u{e015a}: '󠅚'

VariationSelectorDash108

\u{e015b}: '󠅛'

VariationSelectorDash109

\u{e015c}: '󠅜'

VariationSelectorDash110

\u{e015d}: '󠅝'

VariationSelectorDash111

\u{e015e}: '󠅞'

VariationSelectorDash112

\u{e015f}: '󠅟'

VariationSelectorDash113

\u{e0160}: '󠅠'

VariationSelectorDash114

\u{e0161}: '󠅡'

VariationSelectorDash115

\u{e0162}: '󠅢'

VariationSelectorDash116

\u{e0163}: '󠅣'

VariationSelectorDash117

\u{e0164}: '󠅤'

VariationSelectorDash118

\u{e0165}: '󠅥'

VariationSelectorDash119

\u{e0166}: '󠅦'

VariationSelectorDash120

\u{e0167}: '󠅧'

VariationSelectorDash121

\u{e0168}: '󠅨'

VariationSelectorDash122

\u{e0169}: '󠅩'

VariationSelectorDash123

\u{e016a}: '󠅪'

VariationSelectorDash124

\u{e016b}: '󠅫'

VariationSelectorDash125

\u{e016c}: '󠅬'

VariationSelectorDash126

\u{e016d}: '󠅭'

VariationSelectorDash127

\u{e016e}: '󠅮'

VariationSelectorDash128

\u{e016f}: '󠅯'

VariationSelectorDash129

\u{e0170}: '󠅰'

VariationSelectorDash130

\u{e0171}: '󠅱'

VariationSelectorDash131

\u{e0172}: '󠅲'

VariationSelectorDash132

\u{e0173}: '󠅳'

VariationSelectorDash133

\u{e0174}: '󠅴'

VariationSelectorDash134

\u{e0175}: '󠅵'

VariationSelectorDash135

\u{e0176}: '󠅶'

VariationSelectorDash136

\u{e0177}: '󠅷'

VariationSelectorDash137

\u{e0178}: '󠅸'

VariationSelectorDash138

\u{e0179}: '󠅹'

VariationSelectorDash139

\u{e017a}: '󠅺'

VariationSelectorDash140

\u{e017b}: '󠅻'

VariationSelectorDash141

\u{e017c}: '󠅼'

VariationSelectorDash142

\u{e017d}: '󠅽'

VariationSelectorDash143

\u{e017e}: '󠅾'

VariationSelectorDash144

\u{e017f}: '󠅿'

VariationSelectorDash145

\u{e0180}: '󠆀'

VariationSelectorDash146

\u{e0181}: '󠆁'

VariationSelectorDash147

\u{e0182}: '󠆂'

VariationSelectorDash148

\u{e0183}: '󠆃'

VariationSelectorDash149

\u{e0184}: '󠆄'

VariationSelectorDash150

\u{e0185}: '󠆅'

VariationSelectorDash151

\u{e0186}: '󠆆'

VariationSelectorDash152

\u{e0187}: '󠆇'

VariationSelectorDash153

\u{e0188}: '󠆈'

VariationSelectorDash154

\u{e0189}: '󠆉'

VariationSelectorDash155

\u{e018a}: '󠆊'

VariationSelectorDash156

\u{e018b}: '󠆋'

VariationSelectorDash157

\u{e018c}: '󠆌'

VariationSelectorDash158

\u{e018d}: '󠆍'

VariationSelectorDash159

\u{e018e}: '󠆎'

VariationSelectorDash160

\u{e018f}: '󠆏'

VariationSelectorDash161

\u{e0190}: '󠆐'

VariationSelectorDash162

\u{e0191}: '󠆑'

VariationSelectorDash163

\u{e0192}: '󠆒'

VariationSelectorDash164

\u{e0193}: '󠆓'

VariationSelectorDash165

\u{e0194}: '󠆔'

VariationSelectorDash166

\u{e0195}: '󠆕'

VariationSelectorDash167

\u{e0196}: '󠆖'

VariationSelectorDash168

\u{e0197}: '󠆗'

VariationSelectorDash169

\u{e0198}: '󠆘'

VariationSelectorDash170

\u{e0199}: '󠆙'

VariationSelectorDash171

\u{e019a}: '󠆚'

VariationSelectorDash172

\u{e019b}: '󠆛'

VariationSelectorDash173

\u{e019c}: '󠆜'

VariationSelectorDash174

\u{e019d}: '󠆝'

VariationSelectorDash175

\u{e019e}: '󠆞'

VariationSelectorDash176

\u{e019f}: '󠆟'

VariationSelectorDash177

\u{e01a0}: '󠆠'

VariationSelectorDash178

\u{e01a1}: '󠆡'

VariationSelectorDash179

\u{e01a2}: '󠆢'

VariationSelectorDash180

\u{e01a3}: '󠆣'

VariationSelectorDash181

\u{e01a4}: '󠆤'

VariationSelectorDash182

\u{e01a5}: '󠆥'

VariationSelectorDash183

\u{e01a6}: '󠆦'

VariationSelectorDash184

\u{e01a7}: '󠆧'

VariationSelectorDash185

\u{e01a8}: '󠆨'

VariationSelectorDash186

\u{e01a9}: '󠆩'

VariationSelectorDash187

\u{e01aa}: '󠆪'

VariationSelectorDash188

\u{e01ab}: '󠆫'

VariationSelectorDash189

\u{e01ac}: '󠆬'

VariationSelectorDash190

\u{e01ad}: '󠆭'

VariationSelectorDash191

\u{e01ae}: '󠆮'

VariationSelectorDash192

\u{e01af}: '󠆯'

VariationSelectorDash193

\u{e01b0}: '󠆰'

VariationSelectorDash194

\u{e01b1}: '󠆱'

VariationSelectorDash195

\u{e01b2}: '󠆲'

VariationSelectorDash196

\u{e01b3}: '󠆳'

VariationSelectorDash197

\u{e01b4}: '󠆴'

VariationSelectorDash198

\u{e01b5}: '󠆵'

VariationSelectorDash199

\u{e01b6}: '󠆶'

VariationSelectorDash200

\u{e01b7}: '󠆷'

VariationSelectorDash201

\u{e01b8}: '󠆸'

VariationSelectorDash202

\u{e01b9}: '󠆹'

VariationSelectorDash203

\u{e01ba}: '󠆺'

VariationSelectorDash204

\u{e01bb}: '󠆻'

VariationSelectorDash205

\u{e01bc}: '󠆼'

VariationSelectorDash206

\u{e01bd}: '󠆽'

VariationSelectorDash207

\u{e01be}: '󠆾'

VariationSelectorDash208

\u{e01bf}: '󠆿'

VariationSelectorDash209

\u{e01c0}: '󠇀'

VariationSelectorDash210

\u{e01c1}: '󠇁'

VariationSelectorDash211

\u{e01c2}: '󠇂'

VariationSelectorDash212

\u{e01c3}: '󠇃'

VariationSelectorDash213

\u{e01c4}: '󠇄'

VariationSelectorDash214

\u{e01c5}: '󠇅'

VariationSelectorDash215

\u{e01c6}: '󠇆'

VariationSelectorDash216

\u{e01c7}: '󠇇'

VariationSelectorDash217

\u{e01c8}: '󠇈'

VariationSelectorDash218

\u{e01c9}: '󠇉'

VariationSelectorDash219

\u{e01ca}: '󠇊'

VariationSelectorDash220

\u{e01cb}: '󠇋'

VariationSelectorDash221

\u{e01cc}: '󠇌'

VariationSelectorDash222

\u{e01cd}: '󠇍'

VariationSelectorDash223

\u{e01ce}: '󠇎'

VariationSelectorDash224

\u{e01cf}: '󠇏'

VariationSelectorDash225

\u{e01d0}: '󠇐'

VariationSelectorDash226

\u{e01d1}: '󠇑'

VariationSelectorDash227

\u{e01d2}: '󠇒'

VariationSelectorDash228

\u{e01d3}: '󠇓'

VariationSelectorDash229

\u{e01d4}: '󠇔'

VariationSelectorDash230

\u{e01d5}: '󠇕'

VariationSelectorDash231

\u{e01d6}: '󠇖'

VariationSelectorDash232

\u{e01d7}: '󠇗'

VariationSelectorDash233

\u{e01d8}: '󠇘'

VariationSelectorDash234

\u{e01d9}: '󠇙'

VariationSelectorDash235

\u{e01da}: '󠇚'

VariationSelectorDash236

\u{e01db}: '󠇛'

VariationSelectorDash237

\u{e01dc}: '󠇜'

VariationSelectorDash238

\u{e01dd}: '󠇝'

VariationSelectorDash239

\u{e01de}: '󠇞'

VariationSelectorDash240

\u{e01df}: '󠇟'

VariationSelectorDash241

\u{e01e0}: '󠇠'

VariationSelectorDash242

\u{e01e1}: '󠇡'

VariationSelectorDash243

\u{e01e2}: '󠇢'

VariationSelectorDash244

\u{e01e3}: '󠇣'

VariationSelectorDash245

\u{e01e4}: '󠇤'

VariationSelectorDash246

\u{e01e5}: '󠇥'

VariationSelectorDash247

\u{e01e6}: '󠇦'

VariationSelectorDash248

\u{e01e7}: '󠇧'

VariationSelectorDash249

\u{e01e8}: '󠇨'

VariationSelectorDash250

\u{e01e9}: '󠇩'

VariationSelectorDash251

\u{e01ea}: '󠇪'

VariationSelectorDash252

\u{e01eb}: '󠇫'

VariationSelectorDash253

\u{e01ec}: '󠇬'

VariationSelectorDash254

\u{e01ed}: '󠇭'

VariationSelectorDash255

\u{e01ee}: '󠇮'

Methods

impl VariationSelectorsSupplement

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case