Enum unicode_types::warang_citi::WarangCiti

pub enum WarangCiti {
    CapitalLetterNgaa,
    CapitalLetterA,
    CapitalLetterWi,
    CapitalLetterYu,
    CapitalLetterYa,
    CapitalLetterYo,
    CapitalLetterIi,
    CapitalLetterUu,
    CapitalLetterE,
    CapitalLetterO,
    CapitalLetterAng,
    CapitalLetterGa,
    CapitalLetterKo,
    CapitalLetterEny,
    CapitalLetterYuj,
    CapitalLetterUc,
    CapitalLetterEnn,
    CapitalLetterOdd,
    CapitalLetterTte,
    CapitalLetterNung,
    CapitalLetterDa,
    CapitalLetterAt,
    CapitalLetterAm,
    CapitalLetterBu,
    CapitalLetterPu,
    CapitalLetterHiyo,
    CapitalLetterHolo,
    CapitalLetterHorr,
    CapitalLetterHar,
    CapitalLetterSsuu,
    CapitalLetterSii,
    CapitalLetterViyo,
    SmallLetterNgaa,
    SmallLetterA,
    SmallLetterWi,
    SmallLetterYu,
    SmallLetterYa,
    SmallLetterYo,
    SmallLetterIi,
    SmallLetterUu,
    SmallLetterE,
    SmallLetterO,
    SmallLetterAng,
    SmallLetterGa,
    SmallLetterKo,
    SmallLetterEny,
    SmallLetterYuj,
    SmallLetterUc,
    SmallLetterEnn,
    SmallLetterOdd,
    SmallLetterTte,
    SmallLetterNung,
    SmallLetterDa,
    SmallLetterAt,
    SmallLetterAm,
    SmallLetterBu,
    SmallLetterPu,
    SmallLetterHiyo,
    SmallLetterHolo,
    SmallLetterHorr,
    SmallLetterHar,
    SmallLetterSsuu,
    SmallLetterSii,
    SmallLetterViyo,
    DigitZero,
    DigitOne,
    DigitTwo,
    DigitThree,
    DigitFour,
    DigitFive,
    DigitSix,
    DigitSeven,
    DigitEight,
    DigitNine,
    NumberTen,
    NumberTwenty,
    NumberThirty,
    NumberForty,
    NumberFifty,
    NumberSixty,
    NumberSeventy,
    NumberEighty,
    NumberNinety,
}

An enum to represent all characters in the WarangCiti block.

Variants
CapitalLetterNgaa

\u{118a0}: 'ğ‘¢ '

CapitalLetterA

\u{118a1}: 'ğ‘¢¡'

CapitalLetterWi

\u{118a2}: 'ğ‘¢¢'

CapitalLetterYu

\u{118a3}: 'ğ‘¢£'

CapitalLetterYa

\u{118a4}: 'ğ‘¢¤'

CapitalLetterYo

\u{118a5}: 'ğ‘¢¥'

CapitalLetterIi

\u{118a6}: 'ğ‘¢¦'

CapitalLetterUu

\u{118a7}: 'ğ‘¢'

CapitalLetterE

\u{118a8}: 'ğ‘¢¨'

CapitalLetterO

\u{118a9}: 'ğ‘¢©'

CapitalLetterAng

\u{118aa}: 'ğ‘¢ª'

CapitalLetterGa

\u{118ab}: 'ğ‘¢«'

CapitalLetterKo

\u{118ac}: 'ğ‘¢¬'

CapitalLetterEny

\u{118ad}: 'ğ‘¢­'

CapitalLetterYuj

\u{118ae}: 'ğ‘¢®'

CapitalLetterUc

\u{118af}: 'ğ‘¢¯'

CapitalLetterEnn

\u{118b0}: 'ğ‘¢°'

CapitalLetterOdd

\u{118b1}: 'ğ‘¢±'

CapitalLetterTte

\u{118b2}: 'ğ‘¢²'

CapitalLetterNung

\u{118b3}: 'ğ‘¢³'

CapitalLetterDa

\u{118b4}: 'ğ‘¢´'

CapitalLetterAt

\u{118b5}: 'ğ‘¢µ'

CapitalLetterAm

\u{118b6}: 'ğ‘¢¶'

CapitalLetterBu

\u{118b7}: 'ğ‘¢·'

CapitalLetterPu

\u{118b8}: 'ğ‘¢¸'

CapitalLetterHiyo

\u{118b9}: 'ğ‘¢¹'

CapitalLetterHolo

\u{118ba}: 'ğ‘¢º'

CapitalLetterHorr

\u{118bb}: 'ğ‘¢»'

CapitalLetterHar

\u{118bc}: 'ğ‘¢¼'

CapitalLetterSsuu

\u{118bd}: 'ğ‘¢½'

CapitalLetterSii

\u{118be}: 'ğ‘¢¾'

CapitalLetterViyo

\u{118bf}: 'ğ‘¢¿'

SmallLetterNgaa

\u{118c0}: 'ğ‘£€'

SmallLetterA

\u{118c1}: 'ğ‘£'

SmallLetterWi

\u{118c2}: 'ğ‘£‚'

SmallLetterYu

\u{118c3}: 'ğ‘£ƒ'

SmallLetterYa

\u{118c4}: 'ğ‘£„'

SmallLetterYo

\u{118c5}: 'ğ‘£…'

SmallLetterIi

\u{118c6}: 'ğ‘£†'

SmallLetterUu

\u{118c7}: 'ğ‘£‡'

SmallLetterE

\u{118c8}: 'ğ‘£ˆ'

SmallLetterO

\u{118c9}: 'ğ‘£‰'

SmallLetterAng

\u{118ca}: 'ğ‘£Š'

SmallLetterGa

\u{118cb}: 'ğ‘£‹'

SmallLetterKo

\u{118cc}: 'ğ‘£Œ'

SmallLetterEny

\u{118cd}: 'ğ‘£'

SmallLetterYuj

\u{118ce}: 'ğ‘£'

SmallLetterUc

\u{118cf}: 'ğ‘£'

SmallLetterEnn

\u{118d0}: 'ğ‘£'

SmallLetterOdd

\u{118d1}: 'ğ‘£‘'

SmallLetterTte

\u{118d2}: 'ğ‘£’'

SmallLetterNung

\u{118d3}: 'ğ‘£“'

SmallLetterDa

\u{118d4}: 'ğ‘£”'

SmallLetterAt

\u{118d5}: 'ğ‘£•'

SmallLetterAm

\u{118d6}: 'ğ‘£–'

SmallLetterBu

\u{118d7}: 'ğ‘£—'

SmallLetterPu

\u{118d8}: 'ğ‘£˜'

SmallLetterHiyo

\u{118d9}: 'ğ‘£™'

SmallLetterHolo

\u{118da}: 'ğ‘£š'

SmallLetterHorr

\u{118db}: 'ğ‘£›'

SmallLetterHar

\u{118dc}: 'ğ‘£œ'

SmallLetterSsuu

\u{118dd}: 'ğ‘£'

SmallLetterSii

\u{118de}: 'ğ‘£'

SmallLetterViyo

\u{118df}: 'ğ‘£Ÿ'

DigitZero

\u{118e0}: 'ğ‘£ '

DigitOne

\u{118e1}: 'ğ‘£¡'

DigitTwo

\u{118e2}: 'ğ‘£¢'

DigitThree

\u{118e3}: 'ğ‘££'

DigitFour

\u{118e4}: 'ğ‘£¤'

DigitFive

\u{118e5}: 'ğ‘£¥'

DigitSix

\u{118e6}: 'ğ‘£¦'

DigitSeven

\u{118e7}: 'ğ‘£'

DigitEight

\u{118e8}: 'ğ‘£¨'

DigitNine

\u{118e9}: 'ğ‘£©'

NumberTen

\u{118ea}: 'ğ‘£ª'

NumberTwenty

\u{118eb}: 'ğ‘£«'

NumberThirty

\u{118ec}: 'ğ‘£¬'

NumberForty

\u{118ed}: 'ğ‘£­'

NumberFifty

\u{118ee}: 'ğ‘£®'

NumberSixty

\u{118ef}: 'ğ‘£¯'

NumberSeventy

\u{118f0}: 'ğ‘£°'

NumberEighty

\u{118f1}: 'ğ‘£±'

NumberNinety

\u{118f2}: 'ğ‘£²'

Methods

impl WarangCiti

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case