Enum unicode_types::yi_syllables::YiSyllables

pub enum YiSyllables {
    YiSyllableIt,
    YiSyllableIx,
    YiSyllableI,
    YiSyllableIp,
    YiSyllableIet,
    YiSyllableIex,
    YiSyllableIe,
    YiSyllableIep,
    YiSyllableAt,
    YiSyllableAx,
    YiSyllableA,
    YiSyllableAp,
    YiSyllableUox,
    YiSyllableUo,
    YiSyllableUop,
    YiSyllableOt,
    YiSyllableOx,
    YiSyllableO,
    YiSyllableOp,
    YiSyllableEx,
    YiSyllableE,
    YiSyllableWu,
    YiSyllableBit,
    YiSyllableBix,
    YiSyllableBi,
    YiSyllableBip,
    YiSyllableBiet,
    YiSyllableBiex,
    YiSyllableBie,
    YiSyllableBiep,
    YiSyllableBat,
    YiSyllableBax,
    YiSyllableBa,
    YiSyllableBap,
    YiSyllableBuox,
    YiSyllableBuo,
    YiSyllableBuop,
    YiSyllableBot,
    YiSyllableBox,
    YiSyllableBo,
    YiSyllableBop,
    YiSyllableBex,
    YiSyllableBe,
    YiSyllableBep,
    YiSyllableBut,
    YiSyllableBux,
    YiSyllableBu,
    YiSyllableBup,
    YiSyllableBurx,
    YiSyllableBur,
    YiSyllableByt,
    YiSyllableByx,
    YiSyllableBy,
    YiSyllableByp,
    YiSyllableByrx,
    YiSyllableByr,
    YiSyllablePit,
    YiSyllablePix,
    YiSyllablePi,
    YiSyllablePip,
    YiSyllablePiex,
    YiSyllablePie,
    YiSyllablePiep,
    YiSyllablePat,
    YiSyllablePax,
    YiSyllablePa,
    YiSyllablePap,
    YiSyllablePuox,
    YiSyllablePuo,
    YiSyllablePuop,
    YiSyllablePot,
    YiSyllablePox,
    YiSyllablePo,
    YiSyllablePop,
    YiSyllablePut,
    YiSyllablePux,
    YiSyllablePu,
    YiSyllablePup,
    YiSyllablePurx,
    YiSyllablePur,
    YiSyllablePyt,
    YiSyllablePyx,
    YiSyllablePy,
    YiSyllablePyp,
    YiSyllablePyrx,
    YiSyllablePyr,
    YiSyllableBbit,
    YiSyllableBbix,
    YiSyllableBbi,
    YiSyllableBbip,
    YiSyllableBbiet,
    YiSyllableBbiex,
    YiSyllableBbie,
    YiSyllableBbiep,
    YiSyllableBbat,
    YiSyllableBbax,
    YiSyllableBba,
    YiSyllableBbap,
    YiSyllableBbuox,
    YiSyllableBbuo,
    YiSyllableBbuop,
    YiSyllableBbot,
    YiSyllableBbox,
    YiSyllableBbo,
    YiSyllableBbop,
    YiSyllableBbex,
    YiSyllableBbe,
    YiSyllableBbep,
    YiSyllableBbut,
    YiSyllableBbux,
    YiSyllableBbu,
    YiSyllableBbup,
    YiSyllableBburx,
    YiSyllableBbur,
    YiSyllableBbyt,
    YiSyllableBbyx,
    YiSyllableBby,
    YiSyllableBbyp,
    YiSyllableNbit,
    YiSyllableNbix,
    YiSyllableNbi,
    YiSyllableNbip,
    YiSyllableNbiex,
    YiSyllableNbie,
    YiSyllableNbiep,
    YiSyllableNbat,
    YiSyllableNbax,
    YiSyllableNba,
    YiSyllableNbap,
    YiSyllableNbot,
    YiSyllableNbox,
    YiSyllableNbo,
    YiSyllableNbop,
    YiSyllableNbut,
    YiSyllableNbux,
    YiSyllableNbu,
    YiSyllableNbup,
    YiSyllableNburx,
    YiSyllableNbur,
    YiSyllableNbyt,
    YiSyllableNbyx,
    YiSyllableNby,
    YiSyllableNbyp,
    YiSyllableNbyrx,
    YiSyllableNbyr,
    YiSyllableHmit,
    YiSyllableHmix,
    YiSyllableHmi,
    YiSyllableHmip,
    YiSyllableHmiex,
    YiSyllableHmie,
    YiSyllableHmiep,
    YiSyllableHmat,
    YiSyllableHmax,
    YiSyllableHma,
    YiSyllableHmap,
    YiSyllableHmuox,
    YiSyllableHmuo,
    YiSyllableHmuop,
    YiSyllableHmot,
    YiSyllableHmox,
    YiSyllableHmo,
    YiSyllableHmop,
    YiSyllableHmut,
    YiSyllableHmux,
    YiSyllableHmu,
    YiSyllableHmup,
    YiSyllableHmurx,
    YiSyllableHmur,
    YiSyllableHmyx,
    YiSyllableHmy,
    YiSyllableHmyp,
    YiSyllableHmyrx,
    YiSyllableHmyr,
    YiSyllableMit,
    YiSyllableMix,
    YiSyllableMi,
    YiSyllableMip,
    YiSyllableMiex,
    YiSyllableMie,
    YiSyllableMiep,
    YiSyllableMat,
    YiSyllableMax,
    YiSyllableMa,
    YiSyllableMap,
    YiSyllableMuot,
    YiSyllableMuox,
    YiSyllableMuo,
    YiSyllableMuop,
    YiSyllableMot,
    YiSyllableMox,
    YiSyllableMo,
    YiSyllableMop,
    YiSyllableMex,
    YiSyllableMe,
    YiSyllableMut,
    YiSyllableMux,
    YiSyllableMu,
    YiSyllableMup,
    YiSyllableMurx,
    YiSyllableMur,
    YiSyllableMyt,
    YiSyllableMyx,
    YiSyllableMy,
    YiSyllableMyp,
    YiSyllableFit,
    YiSyllableFix,
    YiSyllableFi,
    YiSyllableFip,
    YiSyllableFat,
    YiSyllableFax,
    YiSyllableFa,
    YiSyllableFap,
    YiSyllableFox,
    YiSyllableFo,
    YiSyllableFop,
    YiSyllableFut,
    YiSyllableFux,
    YiSyllableFu,
    YiSyllableFup,
    YiSyllableFurx,
    YiSyllableFur,
    YiSyllableFyt,
    YiSyllableFyx,
    YiSyllableFy,
    YiSyllableFyp,
    YiSyllableVit,
    YiSyllableVix,
    YiSyllableVi,
    YiSyllableVip,
    YiSyllableViet,
    YiSyllableViex,
    YiSyllableVie,
    YiSyllableViep,
    YiSyllableVat,
    YiSyllableVax,
    YiSyllableVa,
    YiSyllableVap,
    YiSyllableVot,
    YiSyllableVox,
    YiSyllableVo,
    YiSyllableVop,
    YiSyllableVex,
    YiSyllableVep,
    YiSyllableVut,
    YiSyllableVux,
    YiSyllableVu,
    YiSyllableVup,
    YiSyllableVurx,
    YiSyllableVur,
    YiSyllableVyt,
    YiSyllableVyx,
    YiSyllableVy,
    YiSyllableVyp,
    YiSyllableVyrx,
    YiSyllableVyr,
    YiSyllableDit,
    YiSyllableDix,
    YiSyllableDi,
    YiSyllableDip,
    YiSyllableDiex,
    YiSyllableDie,
    YiSyllableDiep,
    YiSyllableDat,
    YiSyllableDax,
    YiSyllableDa,
    YiSyllableDap,
    YiSyllableDuox,
    YiSyllableDuo,
    YiSyllableDot,
    YiSyllableDox,
    YiSyllableDo,
    YiSyllableDop,
    YiSyllableDex,
    YiSyllableDe,
    YiSyllableDep,
    YiSyllableDut,
    YiSyllableDux,
    YiSyllableDu,
    YiSyllableDup,
    YiSyllableDurx,
    YiSyllableDur,
    YiSyllableTit,
    YiSyllableTix,
    YiSyllableTi,
    YiSyllableTip,
    YiSyllableTiex,
    YiSyllableTie,
    YiSyllableTiep,
    YiSyllableTat,
    YiSyllableTax,
    YiSyllableTa,
    YiSyllableTap,
    YiSyllableTuot,
    YiSyllableTuox,
    YiSyllableTuo,
    YiSyllableTuop,
    YiSyllableTot,
    YiSyllableTox,
    YiSyllableTo,
    YiSyllableTop,
    YiSyllableTex,
    YiSyllableTe,
    YiSyllableTep,
    YiSyllableTut,
    YiSyllableTux,
    YiSyllableTu,
    YiSyllableTup,
    YiSyllableTurx,
    YiSyllableTur,
    YiSyllableDdit,
    YiSyllableDdix,
    YiSyllableDdi,
    YiSyllableDdip,
    YiSyllableDdiex,
    YiSyllableDdie,
    YiSyllableDdiep,
    YiSyllableDdat,
    YiSyllableDdax,
    YiSyllableDda,
    YiSyllableDdap,
    YiSyllableDduox,
    YiSyllableDduo,
    YiSyllableDduop,
    YiSyllableDdot,
    YiSyllableDdox,
    YiSyllableDdo,
    YiSyllableDdop,
    YiSyllableDdex,
    YiSyllableDde,
    YiSyllableDdep,
    YiSyllableDdut,
    YiSyllableDdux,
    YiSyllableDdu,
    YiSyllableDdup,
    YiSyllableDdurx,
    YiSyllableDdur,
    YiSyllableNdit,
    YiSyllableNdix,
    YiSyllableNdi,
    YiSyllableNdip,
    YiSyllableNdiex,
    YiSyllableNdie,
    YiSyllableNdat,
    YiSyllableNdax,
    YiSyllableNda,
    YiSyllableNdap,
    YiSyllableNdot,
    YiSyllableNdox,
    YiSyllableNdo,
    YiSyllableNdop,
    YiSyllableNdex,
    YiSyllableNde,
    YiSyllableNdep,
    YiSyllableNdut,
    YiSyllableNdux,
    YiSyllableNdu,
    YiSyllableNdup,
    YiSyllableNdurx,
    YiSyllableNdur,
    YiSyllableHnit,
    YiSyllableHnix,
    YiSyllableHni,
    YiSyllableHnip,
    YiSyllableHniet,
    YiSyllableHniex,
    YiSyllableHnie,
    YiSyllableHniep,
    YiSyllableHnat,
    YiSyllableHnax,
    YiSyllableHna,
    YiSyllableHnap,
    YiSyllableHnuox,
    YiSyllableHnuo,
    YiSyllableHnot,
    YiSyllableHnox,
    YiSyllableHnop,
    YiSyllableHnex,
    YiSyllableHne,
    YiSyllableHnep,
    YiSyllableHnut,
    YiSyllableNit,
    YiSyllableNix,
    YiSyllableNi,
    YiSyllableNip,
    YiSyllableNiex,
    YiSyllableNie,
    YiSyllableNiep,
    YiSyllableNax,
    YiSyllableNa,
    YiSyllableNap,
    YiSyllableNuox,
    YiSyllableNuo,
    YiSyllableNuop,
    YiSyllableNot,
    YiSyllableNox,
    YiSyllableNo,
    YiSyllableNop,
    YiSyllableNex,
    YiSyllableNe,
    YiSyllableNep,
    YiSyllableNut,
    YiSyllableNux,
    YiSyllableNu,
    YiSyllableNup,
    YiSyllableNurx,
    YiSyllableNur,
    YiSyllableHlit,
    YiSyllableHlix,
    YiSyllableHli,
    YiSyllableHlip,
    YiSyllableHliex,
    YiSyllableHlie,
    YiSyllableHliep,
    YiSyllableHlat,
    YiSyllableHlax,
    YiSyllableHla,
    YiSyllableHlap,
    YiSyllableHluox,
    YiSyllableHluo,
    YiSyllableHluop,
    YiSyllableHlox,
    YiSyllableHlo,
    YiSyllableHlop,
    YiSyllableHlex,
    YiSyllableHle,
    YiSyllableHlep,
    YiSyllableHlut,
    YiSyllableHlux,
    YiSyllableHlu,
    YiSyllableHlup,
    YiSyllableHlurx,
    YiSyllableHlur,
    YiSyllableHlyt,
    YiSyllableHlyx,
    YiSyllableHly,
    YiSyllableHlyp,
    YiSyllableHlyrx,
    YiSyllableHlyr,
    YiSyllableLit,
    YiSyllableLix,
    YiSyllableLi,
    YiSyllableLip,
    YiSyllableLiet,
    YiSyllableLiex,
    YiSyllableLie,
    YiSyllableLiep,
    YiSyllableLat,
    YiSyllableLax,
    YiSyllableLa,
    YiSyllableLap,
    YiSyllableLuot,
    YiSyllableLuox,
    YiSyllableLuo,
    YiSyllableLuop,
    YiSyllableLot,
    YiSyllableLox,
    YiSyllableLo,
    YiSyllableLop,
    YiSyllableLex,
    YiSyllableLe,
    YiSyllableLep,
    YiSyllableLut,
    YiSyllableLux,
    YiSyllableLu,
    YiSyllableLup,
    YiSyllableLurx,
    YiSyllableLur,
    YiSyllableLyt,
    YiSyllableLyx,
    YiSyllableLy,
    YiSyllableLyp,
    YiSyllableLyrx,
    YiSyllableLyr,
    YiSyllableGit,
    YiSyllableGix,
    YiSyllableGi,
    YiSyllableGip,
    YiSyllableGiet,
    YiSyllableGiex,
    YiSyllableGie,
    YiSyllableGiep,
    YiSyllableGat,
    YiSyllableGax,
    YiSyllableGa,
    YiSyllableGap,
    YiSyllableGuot,
    YiSyllableGuox,
    YiSyllableGuo,
    YiSyllableGuop,
    YiSyllableGot,
    YiSyllableGox,
    YiSyllableGo,
    YiSyllableGop,
    YiSyllableGet,
    YiSyllableGex,
    YiSyllableGe,
    YiSyllableGep,
    YiSyllableGut,
    YiSyllableGux,
    YiSyllableGu,
    YiSyllableGup,
    YiSyllableGurx,
    YiSyllableGur,
    YiSyllableKit,
    YiSyllableKix,
    YiSyllableKi,
    YiSyllableKip,
    YiSyllableKiex,
    YiSyllableKie,
    YiSyllableKiep,
    YiSyllableKat,
    YiSyllableKax,
    YiSyllableKa,
    YiSyllableKap,
    YiSyllableKuox,
    YiSyllableKuo,
    YiSyllableKuop,
    YiSyllableKot,
    YiSyllableKox,
    YiSyllableKo,
    YiSyllableKop,
    YiSyllableKet,
    YiSyllableKex,
    YiSyllableKe,
    YiSyllableKep,
    YiSyllableKut,
    YiSyllableKux,
    YiSyllableKu,
    YiSyllableKup,
    YiSyllableKurx,
    YiSyllableKur,
    YiSyllableGgit,
    YiSyllableGgix,
    YiSyllableGgi,
    YiSyllableGgiex,
    YiSyllableGgie,
    YiSyllableGgiep,
    YiSyllableGgat,
    YiSyllableGgax,
    YiSyllableGga,
    YiSyllableGgap,
    YiSyllableGguot,
    YiSyllableGguox,
    YiSyllableGguo,
    YiSyllableGguop,
    YiSyllableGgot,
    YiSyllableGgox,
    YiSyllableGgo,
    YiSyllableGgop,
    YiSyllableGget,
    YiSyllableGgex,
    YiSyllableGge,
    YiSyllableGgep,
    YiSyllableGgut,
    YiSyllableGgux,
    YiSyllableGgu,
    YiSyllableGgup,
    YiSyllableGgurx,
    YiSyllableGgur,
    YiSyllableMgiex,
    YiSyllableMgie,
    YiSyllableMgat,
    YiSyllableMgax,
    YiSyllableMga,
    YiSyllableMgap,
    YiSyllableMguox,
    YiSyllableMguo,
    YiSyllableMguop,
    YiSyllableMgot,
    YiSyllableMgox,
    YiSyllableMgo,
    YiSyllableMgop,
    YiSyllableMgex,
    YiSyllableMge,
    YiSyllableMgep,
    YiSyllableMgut,
    YiSyllableMgux,
    YiSyllableMgu,
    YiSyllableMgup,
    YiSyllableMgurx,
    YiSyllableMgur,
    YiSyllableHxit,
    YiSyllableHxix,
    YiSyllableHxi,
    YiSyllableHxip,
    YiSyllableHxiet,
    YiSyllableHxiex,
    YiSyllableHxie,
    YiSyllableHxiep,
    YiSyllableHxat,
    YiSyllableHxax,
    YiSyllableHxa,
    YiSyllableHxap,
    YiSyllableHxuot,
    YiSyllableHxuox,
    YiSyllableHxuo,
    YiSyllableHxuop,
    YiSyllableHxot,
    YiSyllableHxox,
    YiSyllableHxo,
    YiSyllableHxop,
    YiSyllableHxex,
    YiSyllableHxe,
    YiSyllableHxep,
    YiSyllableNgiex,
    YiSyllableNgie,
    YiSyllableNgiep,
    YiSyllableNgat,
    YiSyllableNgax,
    YiSyllableNga,
    YiSyllableNgap,
    YiSyllableNguot,
    YiSyllableNguox,
    YiSyllableNguo,
    YiSyllableNgot,
    YiSyllableNgox,
    YiSyllableNgo,
    YiSyllableNgop,
    YiSyllableNgex,
    YiSyllableNge,
    YiSyllableNgep,
    YiSyllableHit,
    YiSyllableHiex,
    YiSyllableHie,
    YiSyllableHat,
    YiSyllableHax,
    YiSyllableHa,
    YiSyllableHap,
    YiSyllableHuot,
    YiSyllableHuox,
    YiSyllableHuo,
    YiSyllableHuop,
    YiSyllableHot,
    YiSyllableHox,
    YiSyllableHo,
    YiSyllableHop,
    YiSyllableHex,
    YiSyllableHe,
    YiSyllableHep,
    YiSyllableWat,
    YiSyllableWax,
    YiSyllableWa,
    YiSyllableWap,
    YiSyllableWuox,
    YiSyllableWuo,
    YiSyllableWuop,
    YiSyllableWox,
    YiSyllableWo,
    YiSyllableWop,
    YiSyllableWex,
    YiSyllableWe,
    YiSyllableWep,
    YiSyllableZit,
    YiSyllableZix,
    YiSyllableZi,
    YiSyllableZip,
    YiSyllableZiex,
    YiSyllableZie,
    YiSyllableZiep,
    YiSyllableZat,
    YiSyllableZax,
    YiSyllableZa,
    YiSyllableZap,
    YiSyllableZuox,
    YiSyllableZuo,
    YiSyllableZuop,
    YiSyllableZot,
    YiSyllableZox,
    YiSyllableZo,
    YiSyllableZop,
    YiSyllableZex,
    YiSyllableZe,
    YiSyllableZep,
    YiSyllableZut,
    YiSyllableZux,
    YiSyllableZu,
    YiSyllableZup,
    YiSyllableZurx,
    YiSyllableZur,
    YiSyllableZyt,
    YiSyllableZyx,
    YiSyllableZy,
    YiSyllableZyp,
    YiSyllableZyrx,
    YiSyllableZyr,
    YiSyllableCit,
    YiSyllableCix,
    YiSyllableCi,
    YiSyllableCip,
    YiSyllableCiet,
    YiSyllableCiex,
    YiSyllableCie,
    YiSyllableCiep,
    YiSyllableCat,
    YiSyllableCax,
    YiSyllableCa,
    YiSyllableCap,
    YiSyllableCuox,
    YiSyllableCuo,
    YiSyllableCuop,
    YiSyllableCot,
    YiSyllableCox,
    YiSyllableCo,
    YiSyllableCop,
    YiSyllableCex,
    YiSyllableCe,
    YiSyllableCep,
    YiSyllableCut,
    YiSyllableCux,
    YiSyllableCu,
    YiSyllableCup,
    YiSyllableCurx,
    YiSyllableCur,
    YiSyllableCyt,
    YiSyllableCyx,
    YiSyllableCy,
    YiSyllableCyp,
    YiSyllableCyrx,
    YiSyllableCyr,
    YiSyllableZzit,
    YiSyllableZzix,
    YiSyllableZzi,
    YiSyllableZzip,
    YiSyllableZziet,
    YiSyllableZziex,
    YiSyllableZzie,
    YiSyllableZziep,
    YiSyllableZzat,
    YiSyllableZzax,
    YiSyllableZza,
    YiSyllableZzap,
    YiSyllableZzox,
    YiSyllableZzo,
    YiSyllableZzop,
    YiSyllableZzex,
    YiSyllableZze,
    YiSyllableZzep,
    YiSyllableZzux,
    YiSyllableZzu,
    YiSyllableZzup,
    YiSyllableZzurx,
    YiSyllableZzur,
    YiSyllableZzyt,
    YiSyllableZzyx,
    YiSyllableZzy,
    YiSyllableZzyp,
    YiSyllableZzyrx,
    YiSyllableZzyr,
    YiSyllableNzit,
    YiSyllableNzix,
    YiSyllableNzi,
    YiSyllableNzip,
    YiSyllableNziex,
    YiSyllableNzie,
    YiSyllableNziep,
    YiSyllableNzat,
    YiSyllableNzax,
    YiSyllableNza,
    YiSyllableNzap,
    YiSyllableNzuox,
    YiSyllableNzuo,
    YiSyllableNzox,
    YiSyllableNzop,
    YiSyllableNzex,
    YiSyllableNze,
    YiSyllableNzux,
    YiSyllableNzu,
    YiSyllableNzup,
    YiSyllableNzurx,
    YiSyllableNzur,
    YiSyllableNzyt,
    YiSyllableNzyx,
    YiSyllableNzy,
    YiSyllableNzyp,
    YiSyllableNzyrx,
    YiSyllableNzyr,
    YiSyllableSit,
    YiSyllableSix,
    YiSyllableSi,
    YiSyllableSip,
    YiSyllableSiex,
    YiSyllableSie,
    YiSyllableSiep,
    YiSyllableSat,
    YiSyllableSax,
    YiSyllableSa,
    YiSyllableSap,
    YiSyllableSuox,
    YiSyllableSuo,
    YiSyllableSuop,
    YiSyllableSot,
    YiSyllableSox,
    YiSyllableSo,
    YiSyllableSop,
    YiSyllableSex,
    YiSyllableSe,
    YiSyllableSep,
    YiSyllableSut,
    YiSyllableSux,
    YiSyllableSu,
    YiSyllableSup,
    YiSyllableSurx,
    YiSyllableSur,
    YiSyllableSyt,
    YiSyllableSyx,
    YiSyllableSy,
    YiSyllableSyp,
    YiSyllableSyrx,
    YiSyllableSyr,
    YiSyllableSsit,
    YiSyllableSsix,
    YiSyllableSsi,
    YiSyllableSsip,
    YiSyllableSsiex,
    YiSyllableSsie,
    YiSyllableSsiep,
    YiSyllableSsat,
    YiSyllableSsax,
    YiSyllableSsa,
    YiSyllableSsap,
    YiSyllableSsot,
    YiSyllableSsox,
    YiSyllableSso,
    YiSyllableSsop,
    YiSyllableSsex,
    YiSyllableSse,
    YiSyllableSsep,
    YiSyllableSsut,
    YiSyllableSsux,
    YiSyllableSsu,
    YiSyllableSsup,
    YiSyllableSsyt,
    YiSyllableSsyx,
    YiSyllableSsy,
    YiSyllableSsyp,
    YiSyllableSsyrx,
    YiSyllableSsyr,
    YiSyllableZhat,
    YiSyllableZhax,
    YiSyllableZha,
    YiSyllableZhap,
    YiSyllableZhuox,
    YiSyllableZhuo,
    YiSyllableZhuop,
    YiSyllableZhot,
    YiSyllableZhox,
    YiSyllableZho,
    YiSyllableZhop,
    YiSyllableZhet,
    YiSyllableZhex,
    YiSyllableZhe,
    YiSyllableZhep,
    YiSyllableZhut,
    YiSyllableZhux,
    YiSyllableZhu,
    YiSyllableZhup,
    YiSyllableZhurx,
    YiSyllableZhur,
    YiSyllableZhyt,
    YiSyllableZhyx,
    YiSyllableZhy,
    YiSyllableZhyp,
    YiSyllableZhyrx,
    YiSyllableZhyr,
    YiSyllableChat,
    YiSyllableChax,
    YiSyllableCha,
    YiSyllableChap,
    YiSyllableChuot,
    YiSyllableChuox,
    YiSyllableChuo,
    YiSyllableChuop,
    YiSyllableChot,
    YiSyllableChox,
    YiSyllableCho,
    YiSyllableChop,
    YiSyllableChet,
    YiSyllableChex,
    YiSyllableChe,
    YiSyllableChep,
    YiSyllableChux,
    YiSyllableChu,
    YiSyllableChup,
    YiSyllableChurx,
    YiSyllableChur,
    YiSyllableChyt,
    YiSyllableChyx,
    YiSyllableChy,
    YiSyllableChyp,
    YiSyllableChyrx,
    YiSyllableChyr,
    YiSyllableRrax,
    YiSyllableRra,
    YiSyllableRruox,
    YiSyllableRruo,
    YiSyllableRrot,
    YiSyllableRrox,
    YiSyllableRro,
    YiSyllableRrop,
    YiSyllableRret,
    YiSyllableRrex,
    YiSyllableRre,
    YiSyllableRrep,
    YiSyllableRrut,
    YiSyllableRrux,
    YiSyllableRru,
    YiSyllableRrup,
    YiSyllableRrurx,
    YiSyllableRrur,
    YiSyllableRryt,
    YiSyllableRryx,
    YiSyllableRry,
    YiSyllableRryp,
    YiSyllableRryrx,
    YiSyllableRryr,
    YiSyllableNrat,
    YiSyllableNrax,
    YiSyllableNra,
    YiSyllableNrap,
    YiSyllableNrox,
    YiSyllableNro,
    YiSyllableNrop,
    YiSyllableNret,
    YiSyllableNrex,
    YiSyllableNre,
    YiSyllableNrep,
    YiSyllableNrut,
    YiSyllableNrux,
    YiSyllableNru,
    YiSyllableNrup,
    YiSyllableNrurx,
    YiSyllableNrur,
    YiSyllableNryt,
    YiSyllableNryx,
    YiSyllableNry,
    YiSyllableNryp,
    YiSyllableNryrx,
    YiSyllableNryr,
    YiSyllableShat,
    YiSyllableShax,
    YiSyllableSha,
    YiSyllableShap,
    YiSyllableShuox,
    YiSyllableShuo,
    YiSyllableShuop,
    YiSyllableShot,
    YiSyllableShox,
    YiSyllableSho,
    YiSyllableShop,
    YiSyllableShet,
    YiSyllableShex,
    YiSyllableShe,
    YiSyllableShep,
    YiSyllableShut,
    YiSyllableShux,
    YiSyllableShu,
    YiSyllableShup,
    YiSyllableShurx,
    YiSyllableShur,
    YiSyllableShyt,
    YiSyllableShyx,
    YiSyllableShy,
    YiSyllableShyp,
    YiSyllableShyrx,
    YiSyllableShyr,
    YiSyllableRat,
    YiSyllableRax,
    YiSyllableRa,
    YiSyllableRap,
    YiSyllableRuox,
    YiSyllableRuo,
    YiSyllableRuop,
    YiSyllableRot,
    YiSyllableRox,
    YiSyllableRo,
    YiSyllableRop,
    YiSyllableRex,
    YiSyllableRe,
    YiSyllableRep,
    YiSyllableRut,
    YiSyllableRux,
    YiSyllableRu,
    YiSyllableRup,
    YiSyllableRurx,
    YiSyllableRur,
    YiSyllableRyt,
    YiSyllableRyx,
    YiSyllableRy,
    YiSyllableRyp,
    YiSyllableRyrx,
    YiSyllableRyr,
    YiSyllableJit,
    YiSyllableJix,
    YiSyllableJi,
    YiSyllableJip,
    YiSyllableJiet,
    YiSyllableJiex,
    YiSyllableJie,
    YiSyllableJiep,
    YiSyllableJuot,
    YiSyllableJuox,
    YiSyllableJuo,
    YiSyllableJuop,
    YiSyllableJot,
    YiSyllableJox,
    YiSyllableJo,
    YiSyllableJop,
    YiSyllableJut,
    YiSyllableJux,
    YiSyllableJu,
    YiSyllableJup,
    YiSyllableJurx,
    YiSyllableJur,
    YiSyllableJyt,
    YiSyllableJyx,
    YiSyllableJy,
    YiSyllableJyp,
    YiSyllableJyrx,
    YiSyllableJyr,
    YiSyllableQit,
    YiSyllableQix,
    YiSyllableQi,
    YiSyllableQip,
    YiSyllableQiet,
    YiSyllableQiex,
    YiSyllableQie,
    YiSyllableQiep,
    YiSyllableQuot,
    YiSyllableQuox,
    YiSyllableQuo,
    YiSyllableQuop,
    YiSyllableQot,
    YiSyllableQox,
    YiSyllableQo,
    YiSyllableQop,
    YiSyllableQut,
    YiSyllableQux,
    YiSyllableQu,
    YiSyllableQup,
    YiSyllableQurx,
    YiSyllableQur,
    YiSyllableQyt,
    YiSyllableQyx,
    YiSyllableQy,
    YiSyllableQyp,
    YiSyllableQyrx,
    YiSyllableQyr,
    YiSyllableJjit,
    YiSyllableJjix,
    YiSyllableJji,
    YiSyllableJjip,
    YiSyllableJjiet,
    YiSyllableJjiex,
    YiSyllableJjie,
    YiSyllableJjiep,
    YiSyllableJjuox,
    YiSyllableJjuo,
    YiSyllableJjuop,
    YiSyllableJjot,
    YiSyllableJjox,
    YiSyllableJjo,
    YiSyllableJjop,
    YiSyllableJjut,
    YiSyllableJjux,
    YiSyllableJju,
    YiSyllableJjup,
    YiSyllableJjurx,
    YiSyllableJjur,
    YiSyllableJjyt,
    YiSyllableJjyx,
    YiSyllableJjy,
    YiSyllableJjyp,
    YiSyllableNjit,
    YiSyllableNjix,
    YiSyllableNji,
    YiSyllableNjip,
    YiSyllableNjiet,
    YiSyllableNjiex,
    YiSyllableNjie,
    YiSyllableNjiep,
    YiSyllableNjuox,
    YiSyllableNjuo,
    YiSyllableNjot,
    YiSyllableNjox,
    YiSyllableNjo,
    YiSyllableNjop,
    YiSyllableNjux,
    YiSyllableNju,
    YiSyllableNjup,
    YiSyllableNjurx,
    YiSyllableNjur,
    YiSyllableNjyt,
    YiSyllableNjyx,
    YiSyllableNjy,
    YiSyllableNjyp,
    YiSyllableNjyrx,
    YiSyllableNjyr,
    YiSyllableNyit,
    YiSyllableNyix,
    YiSyllableNyi,
    YiSyllableNyip,
    YiSyllableNyiet,
    YiSyllableNyiex,
    YiSyllableNyie,
    YiSyllableNyiep,
    YiSyllableNyuox,
    YiSyllableNyuo,
    YiSyllableNyuop,
    YiSyllableNyot,
    YiSyllableNyox,
    YiSyllableNyo,
    YiSyllableNyop,
    YiSyllableNyut,
    YiSyllableNyux,
    YiSyllableNyu,
    YiSyllableNyup,
    YiSyllableXit,
    YiSyllableXix,
    YiSyllableXi,
    YiSyllableXip,
    YiSyllableXiet,
    YiSyllableXiex,
    YiSyllableXie,
    YiSyllableXiep,
    YiSyllableXuox,
    YiSyllableXuo,
    YiSyllableXot,
    YiSyllableXox,
    YiSyllableXo,
    YiSyllableXop,
    YiSyllableXyt,
    YiSyllableXyx,
    YiSyllableXy,
    YiSyllableXyp,
    YiSyllableXyrx,
    YiSyllableXyr,
    YiSyllableYit,
    YiSyllableYix,
    YiSyllableYi,
    YiSyllableYip,
    YiSyllableYiet,
    YiSyllableYiex,
    YiSyllableYie,
    YiSyllableYiep,
    YiSyllableYuot,
    YiSyllableYuox,
    YiSyllableYuo,
    YiSyllableYuop,
    YiSyllableYot,
    YiSyllableYox,
    YiSyllableYo,
    YiSyllableYop,
    YiSyllableYut,
    YiSyllableYux,
    YiSyllableYu,
    YiSyllableYup,
    YiSyllableYurx,
    YiSyllableYur,
    YiSyllableYyt,
    YiSyllableYyx,
    YiSyllableYy,
    YiSyllableYyp,
    YiSyllableYyrx,
    YiSyllableYyr,
}

An enum to represent all characters in the YiSyllables block.

Variants
YiSyllableIt

\u{a000}: ''

YiSyllableIx

\u{a001}: ''

YiSyllableI

\u{a002}: ''

YiSyllableIp

\u{a003}: ''

YiSyllableIet

\u{a004}: ''

YiSyllableIex

\u{a005}: ''

YiSyllableIe

\u{a006}: ''

YiSyllableIep

\u{a007}: ''

YiSyllableAt

\u{a008}: ''

YiSyllableAx

\u{a009}: ''

YiSyllableA

\u{a00a}: ''

YiSyllableAp

\u{a00b}: ''

YiSyllableUox

\u{a00c}: ''

YiSyllableUo

\u{a00d}: ''

YiSyllableUop

\u{a00e}: ''

YiSyllableOt

\u{a00f}: ''

YiSyllableOx

\u{a010}: ''

YiSyllableO

\u{a011}: ''

YiSyllableOp

\u{a012}: ''

YiSyllableEx

\u{a013}: ''

YiSyllableE

\u{a014}: ''

YiSyllableWu

\u{a015}: ''

YiSyllableBit

\u{a016}: ''

YiSyllableBix

\u{a017}: ''

YiSyllableBi

\u{a018}: ''

YiSyllableBip

\u{a019}: ''

YiSyllableBiet

\u{a01a}: ''

YiSyllableBiex

\u{a01b}: ''

YiSyllableBie

\u{a01c}: ''

YiSyllableBiep

\u{a01d}: ''

YiSyllableBat

\u{a01e}: ''

YiSyllableBax

\u{a01f}: ''

YiSyllableBa

\u{a020}: ''

YiSyllableBap

\u{a021}: ''

YiSyllableBuox

\u{a022}: ''

YiSyllableBuo

\u{a023}: ''

YiSyllableBuop

\u{a024}: ''

YiSyllableBot

\u{a025}: ''

YiSyllableBox

\u{a026}: ''

YiSyllableBo

\u{a027}: ''

YiSyllableBop

\u{a028}: ''

YiSyllableBex

\u{a029}: ''

YiSyllableBe

\u{a02a}: ''

YiSyllableBep

\u{a02b}: ''

YiSyllableBut

\u{a02c}: ''

YiSyllableBux

\u{a02d}: ''

YiSyllableBu

\u{a02e}: ''

YiSyllableBup

\u{a02f}: ''

YiSyllableBurx

\u{a030}: ''

YiSyllableBur

\u{a031}: ''

YiSyllableByt

\u{a032}: ''

YiSyllableByx

\u{a033}: ''

YiSyllableBy

\u{a034}: ''

YiSyllableByp

\u{a035}: ''

YiSyllableByrx

\u{a036}: ''

YiSyllableByr

\u{a037}: ''

YiSyllablePit

\u{a038}: ''

YiSyllablePix

\u{a039}: ''

YiSyllablePi

\u{a03a}: ''

YiSyllablePip

\u{a03b}: ''

YiSyllablePiex

\u{a03c}: ''

YiSyllablePie

\u{a03d}: ''

YiSyllablePiep

\u{a03e}: ''

YiSyllablePat

\u{a03f}: ''

YiSyllablePax

\u{a040}: ''

YiSyllablePa

\u{a041}: ''

YiSyllablePap

\u{a042}: ''

YiSyllablePuox

\u{a043}: ''

YiSyllablePuo

\u{a044}: ''

YiSyllablePuop

\u{a045}: ''

YiSyllablePot

\u{a046}: ''

YiSyllablePox

\u{a047}: ''

YiSyllablePo

\u{a048}: ''

YiSyllablePop

\u{a049}: ''

YiSyllablePut

\u{a04a}: ''

YiSyllablePux

\u{a04b}: ''

YiSyllablePu

\u{a04c}: ''

YiSyllablePup

\u{a04d}: ''

YiSyllablePurx

\u{a04e}: ''

YiSyllablePur

\u{a04f}: ''

YiSyllablePyt

\u{a050}: ''

YiSyllablePyx

\u{a051}: ''

YiSyllablePy

\u{a052}: ''

YiSyllablePyp

\u{a053}: ''

YiSyllablePyrx

\u{a054}: ''

YiSyllablePyr

\u{a055}: ''

YiSyllableBbit

\u{a056}: ''

YiSyllableBbix

\u{a057}: ''

YiSyllableBbi

\u{a058}: ''

YiSyllableBbip

\u{a059}: ''

YiSyllableBbiet

\u{a05a}: ''

YiSyllableBbiex

\u{a05b}: ''

YiSyllableBbie

\u{a05c}: ''

YiSyllableBbiep

\u{a05d}: ''

YiSyllableBbat

\u{a05e}: ''

YiSyllableBbax

\u{a05f}: ''

YiSyllableBba

\u{a060}: ''

YiSyllableBbap

\u{a061}: ''

YiSyllableBbuox

\u{a062}: ''

YiSyllableBbuo

\u{a063}: ''

YiSyllableBbuop

\u{a064}: ''

YiSyllableBbot

\u{a065}: ''

YiSyllableBbox

\u{a066}: ''

YiSyllableBbo

\u{a067}: ''

YiSyllableBbop

\u{a068}: ''

YiSyllableBbex

\u{a069}: ''

YiSyllableBbe

\u{a06a}: ''

YiSyllableBbep

\u{a06b}: ''

YiSyllableBbut

\u{a06c}: ''

YiSyllableBbux

\u{a06d}: ''

YiSyllableBbu

\u{a06e}: ''

YiSyllableBbup

\u{a06f}: ''

YiSyllableBburx

\u{a070}: ''

YiSyllableBbur

\u{a071}: ''

YiSyllableBbyt

\u{a072}: ''

YiSyllableBbyx

\u{a073}: ''

YiSyllableBby

\u{a074}: ''

YiSyllableBbyp

\u{a075}: ''

YiSyllableNbit

\u{a076}: ''

YiSyllableNbix

\u{a077}: ''

YiSyllableNbi

\u{a078}: ''

YiSyllableNbip

\u{a079}: ''

YiSyllableNbiex

\u{a07a}: ''

YiSyllableNbie

\u{a07b}: ''

YiSyllableNbiep

\u{a07c}: ''

YiSyllableNbat

\u{a07d}: ''

YiSyllableNbax

\u{a07e}: ''

YiSyllableNba

\u{a07f}: ''

YiSyllableNbap

\u{a080}: ''

YiSyllableNbot

\u{a081}: ''

YiSyllableNbox

\u{a082}: ''

YiSyllableNbo

\u{a083}: ''

YiSyllableNbop

\u{a084}: ''

YiSyllableNbut

\u{a085}: ''

YiSyllableNbux

\u{a086}: ''

YiSyllableNbu

\u{a087}: ''

YiSyllableNbup

\u{a088}: ''

YiSyllableNburx

\u{a089}: ''

YiSyllableNbur

\u{a08a}: ''

YiSyllableNbyt

\u{a08b}: ''

YiSyllableNbyx

\u{a08c}: ''

YiSyllableNby

\u{a08d}: ''

YiSyllableNbyp

\u{a08e}: ''

YiSyllableNbyrx

\u{a08f}: ''

YiSyllableNbyr

\u{a090}: ''

YiSyllableHmit

\u{a091}: ''

YiSyllableHmix

\u{a092}: ''

YiSyllableHmi

\u{a093}: ''

YiSyllableHmip

\u{a094}: ''

YiSyllableHmiex

\u{a095}: ''

YiSyllableHmie

\u{a096}: ''

YiSyllableHmiep

\u{a097}: ''

YiSyllableHmat

\u{a098}: ''

YiSyllableHmax

\u{a099}: ''

YiSyllableHma

\u{a09a}: ''

YiSyllableHmap

\u{a09b}: ''

YiSyllableHmuox

\u{a09c}: ''

YiSyllableHmuo

\u{a09d}: ''

YiSyllableHmuop

\u{a09e}: ''

YiSyllableHmot

\u{a09f}: ''

YiSyllableHmox

\u{a0a0}: ''

YiSyllableHmo

\u{a0a1}: ''

YiSyllableHmop

\u{a0a2}: ''

YiSyllableHmut

\u{a0a3}: ''

YiSyllableHmux

\u{a0a4}: ''

YiSyllableHmu

\u{a0a5}: ''

YiSyllableHmup

\u{a0a6}: ''

YiSyllableHmurx

\u{a0a7}: ''

YiSyllableHmur

\u{a0a8}: ''

YiSyllableHmyx

\u{a0a9}: ''

YiSyllableHmy

\u{a0aa}: ''

YiSyllableHmyp

\u{a0ab}: ''

YiSyllableHmyrx

\u{a0ac}: ''

YiSyllableHmyr

\u{a0ad}: ''

YiSyllableMit

\u{a0ae}: ''

YiSyllableMix

\u{a0af}: ''

YiSyllableMi

\u{a0b0}: ''

YiSyllableMip

\u{a0b1}: ''

YiSyllableMiex

\u{a0b2}: ''

YiSyllableMie

\u{a0b3}: ''

YiSyllableMiep

\u{a0b4}: ''

YiSyllableMat

\u{a0b5}: ''

YiSyllableMax

\u{a0b6}: ''

YiSyllableMa

\u{a0b7}: ''

YiSyllableMap

\u{a0b8}: ''

YiSyllableMuot

\u{a0b9}: ''

YiSyllableMuox

\u{a0ba}: ''

YiSyllableMuo

\u{a0bb}: ''

YiSyllableMuop

\u{a0bc}: ''

YiSyllableMot

\u{a0bd}: ''

YiSyllableMox

\u{a0be}: ''

YiSyllableMo

\u{a0bf}: ''

YiSyllableMop

\u{a0c0}: ''

YiSyllableMex

\u{a0c1}: ''

YiSyllableMe

\u{a0c2}: ''

YiSyllableMut

\u{a0c3}: ''

YiSyllableMux

\u{a0c4}: ''

YiSyllableMu

\u{a0c5}: ''

YiSyllableMup

\u{a0c6}: ''

YiSyllableMurx

\u{a0c7}: ''

YiSyllableMur

\u{a0c8}: ''

YiSyllableMyt

\u{a0c9}: ''

YiSyllableMyx

\u{a0ca}: ''

YiSyllableMy

\u{a0cb}: ''

YiSyllableMyp

\u{a0cc}: ''

YiSyllableFit

\u{a0cd}: ''

YiSyllableFix

\u{a0ce}: ''

YiSyllableFi

\u{a0cf}: ''

YiSyllableFip

\u{a0d0}: ''

YiSyllableFat

\u{a0d1}: ''

YiSyllableFax

\u{a0d2}: ''

YiSyllableFa

\u{a0d3}: ''

YiSyllableFap

\u{a0d4}: ''

YiSyllableFox

\u{a0d5}: ''

YiSyllableFo

\u{a0d6}: ''

YiSyllableFop

\u{a0d7}: ''

YiSyllableFut

\u{a0d8}: ''

YiSyllableFux

\u{a0d9}: ''

YiSyllableFu

\u{a0da}: ''

YiSyllableFup

\u{a0db}: ''

YiSyllableFurx

\u{a0dc}: ''

YiSyllableFur

\u{a0dd}: ''

YiSyllableFyt

\u{a0de}: ''

YiSyllableFyx

\u{a0df}: ''

YiSyllableFy

\u{a0e0}: ''

YiSyllableFyp

\u{a0e1}: ''

YiSyllableVit

\u{a0e2}: ''

YiSyllableVix

\u{a0e3}: ''

YiSyllableVi

\u{a0e4}: ''

YiSyllableVip

\u{a0e5}: ''

YiSyllableViet

\u{a0e6}: ''

YiSyllableViex

\u{a0e7}: ''

YiSyllableVie

\u{a0e8}: ''

YiSyllableViep

\u{a0e9}: ''

YiSyllableVat

\u{a0ea}: ''

YiSyllableVax

\u{a0eb}: ''

YiSyllableVa

\u{a0ec}: ''

YiSyllableVap

\u{a0ed}: ''

YiSyllableVot

\u{a0ee}: ''

YiSyllableVox

\u{a0ef}: ''

YiSyllableVo

\u{a0f0}: ''

YiSyllableVop

\u{a0f1}: ''

YiSyllableVex

\u{a0f2}: ''

YiSyllableVep

\u{a0f3}: ''

YiSyllableVut

\u{a0f4}: ''

YiSyllableVux

\u{a0f5}: ''

YiSyllableVu

\u{a0f6}: ''

YiSyllableVup

\u{a0f7}: ''

YiSyllableVurx

\u{a0f8}: ''

YiSyllableVur

\u{a0f9}: ''

YiSyllableVyt

\u{a0fa}: ''

YiSyllableVyx

\u{a0fb}: ''

YiSyllableVy

\u{a0fc}: ''

YiSyllableVyp

\u{a0fd}: ''

YiSyllableVyrx

\u{a0fe}: ''

YiSyllableVyr

\u{a0ff}: ''

YiSyllableDit

\u{a100}: ''

YiSyllableDix

\u{a101}: ''

YiSyllableDi

\u{a102}: ''

YiSyllableDip

\u{a103}: ''

YiSyllableDiex

\u{a104}: ''

YiSyllableDie

\u{a105}: ''

YiSyllableDiep

\u{a106}: ''

YiSyllableDat

\u{a107}: ''

YiSyllableDax

\u{a108}: ''

YiSyllableDa

\u{a109}: ''

YiSyllableDap

\u{a10a}: ''

YiSyllableDuox

\u{a10b}: ''

YiSyllableDuo

\u{a10c}: ''

YiSyllableDot

\u{a10d}: ''

YiSyllableDox

\u{a10e}: ''

YiSyllableDo

\u{a10f}: ''

YiSyllableDop

\u{a110}: ''

YiSyllableDex

\u{a111}: ''

YiSyllableDe

\u{a112}: ''

YiSyllableDep

\u{a113}: ''

YiSyllableDut

\u{a114}: ''

YiSyllableDux

\u{a115}: ''

YiSyllableDu

\u{a116}: ''

YiSyllableDup

\u{a117}: ''

YiSyllableDurx

\u{a118}: ''

YiSyllableDur

\u{a119}: ''

YiSyllableTit

\u{a11a}: ''

YiSyllableTix

\u{a11b}: ''

YiSyllableTi

\u{a11c}: ''

YiSyllableTip

\u{a11d}: ''

YiSyllableTiex

\u{a11e}: ''

YiSyllableTie

\u{a11f}: ''

YiSyllableTiep

\u{a120}: ''

YiSyllableTat

\u{a121}: ''

YiSyllableTax

\u{a122}: ''

YiSyllableTa

\u{a123}: ''

YiSyllableTap

\u{a124}: ''

YiSyllableTuot

\u{a125}: ''

YiSyllableTuox

\u{a126}: ''

YiSyllableTuo

\u{a127}: ''

YiSyllableTuop

\u{a128}: ''

YiSyllableTot

\u{a129}: ''

YiSyllableTox

\u{a12a}: ''

YiSyllableTo

\u{a12b}: ''

YiSyllableTop

\u{a12c}: ''

YiSyllableTex

\u{a12d}: ''

YiSyllableTe

\u{a12e}: ''

YiSyllableTep

\u{a12f}: ''

YiSyllableTut

\u{a130}: ''

YiSyllableTux

\u{a131}: ''

YiSyllableTu

\u{a132}: ''

YiSyllableTup

\u{a133}: ''

YiSyllableTurx

\u{a134}: ''

YiSyllableTur

\u{a135}: ''

YiSyllableDdit

\u{a136}: ''

YiSyllableDdix

\u{a137}: ''

YiSyllableDdi

\u{a138}: ''

YiSyllableDdip

\u{a139}: ''

YiSyllableDdiex

\u{a13a}: ''

YiSyllableDdie

\u{a13b}: ''

YiSyllableDdiep

\u{a13c}: ''

YiSyllableDdat

\u{a13d}: ''

YiSyllableDdax

\u{a13e}: ''

YiSyllableDda

\u{a13f}: ''

YiSyllableDdap

\u{a140}: ''

YiSyllableDduox

\u{a141}: ''

YiSyllableDduo

\u{a142}: ''

YiSyllableDduop

\u{a143}: ''

YiSyllableDdot

\u{a144}: ''

YiSyllableDdox

\u{a145}: ''

YiSyllableDdo

\u{a146}: ''

YiSyllableDdop

\u{a147}: ''

YiSyllableDdex

\u{a148}: ''

YiSyllableDde

\u{a149}: ''

YiSyllableDdep

\u{a14a}: ''

YiSyllableDdut

\u{a14b}: ''

YiSyllableDdux

\u{a14c}: ''

YiSyllableDdu

\u{a14d}: ''

YiSyllableDdup

\u{a14e}: ''

YiSyllableDdurx

\u{a14f}: ''

YiSyllableDdur

\u{a150}: ''

YiSyllableNdit

\u{a151}: ''

YiSyllableNdix

\u{a152}: ''

YiSyllableNdi

\u{a153}: ''

YiSyllableNdip

\u{a154}: ''

YiSyllableNdiex

\u{a155}: ''

YiSyllableNdie

\u{a156}: ''

YiSyllableNdat

\u{a157}: ''

YiSyllableNdax

\u{a158}: ''

YiSyllableNda

\u{a159}: ''

YiSyllableNdap

\u{a15a}: ''

YiSyllableNdot

\u{a15b}: ''

YiSyllableNdox

\u{a15c}: ''

YiSyllableNdo

\u{a15d}: ''

YiSyllableNdop

\u{a15e}: ''

YiSyllableNdex

\u{a15f}: ''

YiSyllableNde

\u{a160}: ''

YiSyllableNdep

\u{a161}: ''

YiSyllableNdut

\u{a162}: ''

YiSyllableNdux

\u{a163}: ''

YiSyllableNdu

\u{a164}: ''

YiSyllableNdup

\u{a165}: ''

YiSyllableNdurx

\u{a166}: ''

YiSyllableNdur

\u{a167}: ''

YiSyllableHnit

\u{a168}: ''

YiSyllableHnix

\u{a169}: ''

YiSyllableHni

\u{a16a}: ''

YiSyllableHnip

\u{a16b}: ''

YiSyllableHniet

\u{a16c}: ''

YiSyllableHniex

\u{a16d}: ''

YiSyllableHnie

\u{a16e}: ''

YiSyllableHniep

\u{a16f}: ''

YiSyllableHnat

\u{a170}: ''

YiSyllableHnax

\u{a171}: ''

YiSyllableHna

\u{a172}: ''

YiSyllableHnap

\u{a173}: ''

YiSyllableHnuox

\u{a174}: ''

YiSyllableHnuo

\u{a175}: ''

YiSyllableHnot

\u{a176}: ''

YiSyllableHnox

\u{a177}: ''

YiSyllableHnop

\u{a178}: ''

YiSyllableHnex

\u{a179}: ''

YiSyllableHne

\u{a17a}: ''

YiSyllableHnep

\u{a17b}: ''

YiSyllableHnut

\u{a17c}: ''

YiSyllableNit

\u{a17d}: ''

YiSyllableNix

\u{a17e}: ''

YiSyllableNi

\u{a17f}: ''

YiSyllableNip

\u{a180}: ''

YiSyllableNiex

\u{a181}: ''

YiSyllableNie

\u{a182}: ''

YiSyllableNiep

\u{a183}: ''

YiSyllableNax

\u{a184}: ''

YiSyllableNa

\u{a185}: ''

YiSyllableNap

\u{a186}: ''

YiSyllableNuox

\u{a187}: ''

YiSyllableNuo

\u{a188}: ''

YiSyllableNuop

\u{a189}: ''

YiSyllableNot

\u{a18a}: ''

YiSyllableNox

\u{a18b}: ''

YiSyllableNo

\u{a18c}: ''

YiSyllableNop

\u{a18d}: ''

YiSyllableNex

\u{a18e}: ''

YiSyllableNe

\u{a18f}: ''

YiSyllableNep

\u{a190}: ''

YiSyllableNut

\u{a191}: ''

YiSyllableNux

\u{a192}: ''

YiSyllableNu

\u{a193}: ''

YiSyllableNup

\u{a194}: ''

YiSyllableNurx

\u{a195}: ''

YiSyllableNur

\u{a196}: ''

YiSyllableHlit

\u{a197}: ''

YiSyllableHlix

\u{a198}: ''

YiSyllableHli

\u{a199}: ''

YiSyllableHlip

\u{a19a}: ''

YiSyllableHliex

\u{a19b}: ''

YiSyllableHlie

\u{a19c}: ''

YiSyllableHliep

\u{a19d}: ''

YiSyllableHlat

\u{a19e}: ''

YiSyllableHlax

\u{a19f}: ''

YiSyllableHla

\u{a1a0}: ''

YiSyllableHlap

\u{a1a1}: ''

YiSyllableHluox

\u{a1a2}: ''

YiSyllableHluo

\u{a1a3}: ''

YiSyllableHluop

\u{a1a4}: ''

YiSyllableHlox

\u{a1a5}: ''

YiSyllableHlo

\u{a1a6}: ''

YiSyllableHlop

\u{a1a7}: ''

YiSyllableHlex

\u{a1a8}: ''

YiSyllableHle

\u{a1a9}: ''

YiSyllableHlep

\u{a1aa}: ''

YiSyllableHlut

\u{a1ab}: ''

YiSyllableHlux

\u{a1ac}: ''

YiSyllableHlu

\u{a1ad}: ''

YiSyllableHlup

\u{a1ae}: ''

YiSyllableHlurx

\u{a1af}: ''

YiSyllableHlur

\u{a1b0}: ''

YiSyllableHlyt

\u{a1b1}: ''

YiSyllableHlyx

\u{a1b2}: ''

YiSyllableHly

\u{a1b3}: ''

YiSyllableHlyp

\u{a1b4}: ''

YiSyllableHlyrx

\u{a1b5}: ''

YiSyllableHlyr

\u{a1b6}: ''

YiSyllableLit

\u{a1b7}: ''

YiSyllableLix

\u{a1b8}: ''

YiSyllableLi

\u{a1b9}: ''

YiSyllableLip

\u{a1ba}: ''

YiSyllableLiet

\u{a1bb}: ''

YiSyllableLiex

\u{a1bc}: ''

YiSyllableLie

\u{a1bd}: ''

YiSyllableLiep

\u{a1be}: ''

YiSyllableLat

\u{a1bf}: ''

YiSyllableLax

\u{a1c0}: ''

YiSyllableLa

\u{a1c1}: ''

YiSyllableLap

\u{a1c2}: ''

YiSyllableLuot

\u{a1c3}: ''

YiSyllableLuox

\u{a1c4}: ''

YiSyllableLuo

\u{a1c5}: ''

YiSyllableLuop

\u{a1c6}: ''

YiSyllableLot

\u{a1c7}: ''

YiSyllableLox

\u{a1c8}: ''

YiSyllableLo

\u{a1c9}: ''

YiSyllableLop

\u{a1ca}: ''

YiSyllableLex

\u{a1cb}: ''

YiSyllableLe

\u{a1cc}: ''

YiSyllableLep

\u{a1cd}: ''

YiSyllableLut

\u{a1ce}: ''

YiSyllableLux

\u{a1cf}: ''

YiSyllableLu

\u{a1d0}: ''

YiSyllableLup

\u{a1d1}: ''

YiSyllableLurx

\u{a1d2}: ''

YiSyllableLur

\u{a1d3}: ''

YiSyllableLyt

\u{a1d4}: ''

YiSyllableLyx

\u{a1d5}: ''

YiSyllableLy

\u{a1d6}: ''

YiSyllableLyp

\u{a1d7}: ''

YiSyllableLyrx

\u{a1d8}: ''

YiSyllableLyr

\u{a1d9}: ''

YiSyllableGit

\u{a1da}: ''

YiSyllableGix

\u{a1db}: ''

YiSyllableGi

\u{a1dc}: ''

YiSyllableGip

\u{a1dd}: ''

YiSyllableGiet

\u{a1de}: ''

YiSyllableGiex

\u{a1df}: ''

YiSyllableGie

\u{a1e0}: ''

YiSyllableGiep

\u{a1e1}: ''

YiSyllableGat

\u{a1e2}: ''

YiSyllableGax

\u{a1e3}: ''

YiSyllableGa

\u{a1e4}: ''

YiSyllableGap

\u{a1e5}: ''

YiSyllableGuot

\u{a1e6}: ''

YiSyllableGuox

\u{a1e7}: ''

YiSyllableGuo

\u{a1e8}: ''

YiSyllableGuop

\u{a1e9}: ''

YiSyllableGot

\u{a1ea}: ''

YiSyllableGox

\u{a1eb}: ''

YiSyllableGo

\u{a1ec}: ''

YiSyllableGop

\u{a1ed}: ''

YiSyllableGet

\u{a1ee}: ''

YiSyllableGex

\u{a1ef}: ''

YiSyllableGe

\u{a1f0}: ''

YiSyllableGep

\u{a1f1}: ''

YiSyllableGut

\u{a1f2}: ''

YiSyllableGux

\u{a1f3}: ''

YiSyllableGu

\u{a1f4}: ''

YiSyllableGup

\u{a1f5}: ''

YiSyllableGurx

\u{a1f6}: ''

YiSyllableGur

\u{a1f7}: ''

YiSyllableKit

\u{a1f8}: ''

YiSyllableKix

\u{a1f9}: ''

YiSyllableKi

\u{a1fa}: ''

YiSyllableKip

\u{a1fb}: ''

YiSyllableKiex

\u{a1fc}: ''

YiSyllableKie

\u{a1fd}: ''

YiSyllableKiep

\u{a1fe}: ''

YiSyllableKat

\u{a1ff}: ''

YiSyllableKax

\u{a200}: ''

YiSyllableKa

\u{a201}: ''

YiSyllableKap

\u{a202}: ''

YiSyllableKuox

\u{a203}: ''

YiSyllableKuo

\u{a204}: ''

YiSyllableKuop

\u{a205}: ''

YiSyllableKot

\u{a206}: ''

YiSyllableKox

\u{a207}: ''

YiSyllableKo

\u{a208}: ''

YiSyllableKop

\u{a209}: ''

YiSyllableKet

\u{a20a}: ''

YiSyllableKex

\u{a20b}: ''

YiSyllableKe

\u{a20c}: ''

YiSyllableKep

\u{a20d}: ''

YiSyllableKut

\u{a20e}: ''

YiSyllableKux

\u{a20f}: ''

YiSyllableKu

\u{a210}: ''

YiSyllableKup

\u{a211}: ''

YiSyllableKurx

\u{a212}: ''

YiSyllableKur

\u{a213}: ''

YiSyllableGgit

\u{a214}: ''

YiSyllableGgix

\u{a215}: ''

YiSyllableGgi

\u{a216}: ''

YiSyllableGgiex

\u{a217}: ''

YiSyllableGgie

\u{a218}: ''

YiSyllableGgiep

\u{a219}: ''

YiSyllableGgat

\u{a21a}: ''

YiSyllableGgax

\u{a21b}: ''

YiSyllableGga

\u{a21c}: ''

YiSyllableGgap

\u{a21d}: ''

YiSyllableGguot

\u{a21e}: ''

YiSyllableGguox

\u{a21f}: ''

YiSyllableGguo

\u{a220}: ''

YiSyllableGguop

\u{a221}: ''

YiSyllableGgot

\u{a222}: ''

YiSyllableGgox

\u{a223}: ''

YiSyllableGgo

\u{a224}: ''

YiSyllableGgop

\u{a225}: ''

YiSyllableGget

\u{a226}: ''

YiSyllableGgex

\u{a227}: ''

YiSyllableGge

\u{a228}: ''

YiSyllableGgep

\u{a229}: ''

YiSyllableGgut

\u{a22a}: ''

YiSyllableGgux

\u{a22b}: ''

YiSyllableGgu

\u{a22c}: ''

YiSyllableGgup

\u{a22d}: ''

YiSyllableGgurx

\u{a22e}: ''

YiSyllableGgur

\u{a22f}: ''

YiSyllableMgiex

\u{a230}: ''

YiSyllableMgie

\u{a231}: ''

YiSyllableMgat

\u{a232}: ''

YiSyllableMgax

\u{a233}: ''

YiSyllableMga

\u{a234}: ''

YiSyllableMgap

\u{a235}: ''

YiSyllableMguox

\u{a236}: ''

YiSyllableMguo

\u{a237}: ''

YiSyllableMguop

\u{a238}: ''

YiSyllableMgot

\u{a239}: ''

YiSyllableMgox

\u{a23a}: ''

YiSyllableMgo

\u{a23b}: ''

YiSyllableMgop

\u{a23c}: ''

YiSyllableMgex

\u{a23d}: ''

YiSyllableMge

\u{a23e}: ''

YiSyllableMgep

\u{a23f}: ''

YiSyllableMgut

\u{a240}: ''

YiSyllableMgux

\u{a241}: ''

YiSyllableMgu

\u{a242}: ''

YiSyllableMgup

\u{a243}: ''

YiSyllableMgurx

\u{a244}: ''

YiSyllableMgur

\u{a245}: ''

YiSyllableHxit

\u{a246}: ''

YiSyllableHxix

\u{a247}: ''

YiSyllableHxi

\u{a248}: ''

YiSyllableHxip

\u{a249}: ''

YiSyllableHxiet

\u{a24a}: ''

YiSyllableHxiex

\u{a24b}: ''

YiSyllableHxie

\u{a24c}: ''

YiSyllableHxiep

\u{a24d}: ''

YiSyllableHxat

\u{a24e}: ''

YiSyllableHxax

\u{a24f}: ''

YiSyllableHxa

\u{a250}: ''

YiSyllableHxap

\u{a251}: ''

YiSyllableHxuot

\u{a252}: ''

YiSyllableHxuox

\u{a253}: ''

YiSyllableHxuo

\u{a254}: ''

YiSyllableHxuop

\u{a255}: ''

YiSyllableHxot

\u{a256}: ''

YiSyllableHxox

\u{a257}: ''

YiSyllableHxo

\u{a258}: ''

YiSyllableHxop

\u{a259}: ''

YiSyllableHxex

\u{a25a}: ''

YiSyllableHxe

\u{a25b}: ''

YiSyllableHxep

\u{a25c}: ''

YiSyllableNgiex

\u{a25d}: ''

YiSyllableNgie

\u{a25e}: ''

YiSyllableNgiep

\u{a25f}: ''

YiSyllableNgat

\u{a260}: ''

YiSyllableNgax

\u{a261}: ''

YiSyllableNga

\u{a262}: ''

YiSyllableNgap

\u{a263}: ''

YiSyllableNguot

\u{a264}: ''

YiSyllableNguox

\u{a265}: ''

YiSyllableNguo

\u{a266}: ''

YiSyllableNgot

\u{a267}: ''

YiSyllableNgox

\u{a268}: ''

YiSyllableNgo

\u{a269}: ''

YiSyllableNgop

\u{a26a}: ''

YiSyllableNgex

\u{a26b}: ''

YiSyllableNge

\u{a26c}: ''

YiSyllableNgep

\u{a26d}: ''

YiSyllableHit

\u{a26e}: ''

YiSyllableHiex

\u{a26f}: ''

YiSyllableHie

\u{a270}: ''

YiSyllableHat

\u{a271}: ''

YiSyllableHax

\u{a272}: ''

YiSyllableHa

\u{a273}: ''

YiSyllableHap

\u{a274}: ''

YiSyllableHuot

\u{a275}: ''

YiSyllableHuox

\u{a276}: ''

YiSyllableHuo

\u{a277}: ''

YiSyllableHuop

\u{a278}: ''

YiSyllableHot

\u{a279}: ''

YiSyllableHox

\u{a27a}: ''

YiSyllableHo

\u{a27b}: ''

YiSyllableHop

\u{a27c}: ''

YiSyllableHex

\u{a27d}: ''

YiSyllableHe

\u{a27e}: ''

YiSyllableHep

\u{a27f}: ''

YiSyllableWat

\u{a280}: ''

YiSyllableWax

\u{a281}: ''

YiSyllableWa

\u{a282}: ''

YiSyllableWap

\u{a283}: ''

YiSyllableWuox

\u{a284}: ''

YiSyllableWuo

\u{a285}: ''

YiSyllableWuop

\u{a286}: ''

YiSyllableWox

\u{a287}: ''

YiSyllableWo

\u{a288}: ''

YiSyllableWop

\u{a289}: ''

YiSyllableWex

\u{a28a}: ''

YiSyllableWe

\u{a28b}: ''

YiSyllableWep

\u{a28c}: ''

YiSyllableZit

\u{a28d}: ''

YiSyllableZix

\u{a28e}: ''

YiSyllableZi

\u{a28f}: ''

YiSyllableZip

\u{a290}: ''

YiSyllableZiex

\u{a291}: ''

YiSyllableZie

\u{a292}: ''

YiSyllableZiep

\u{a293}: ''

YiSyllableZat

\u{a294}: ''

YiSyllableZax

\u{a295}: ''

YiSyllableZa

\u{a296}: ''

YiSyllableZap

\u{a297}: ''

YiSyllableZuox

\u{a298}: ''

YiSyllableZuo

\u{a299}: ''

YiSyllableZuop

\u{a29a}: ''

YiSyllableZot

\u{a29b}: ''

YiSyllableZox

\u{a29c}: ''

YiSyllableZo

\u{a29d}: ''

YiSyllableZop

\u{a29e}: ''

YiSyllableZex

\u{a29f}: ''

YiSyllableZe

\u{a2a0}: ''

YiSyllableZep

\u{a2a1}: ''

YiSyllableZut

\u{a2a2}: ''

YiSyllableZux

\u{a2a3}: ''

YiSyllableZu

\u{a2a4}: ''

YiSyllableZup

\u{a2a5}: ''

YiSyllableZurx

\u{a2a6}: ''

YiSyllableZur

\u{a2a7}: ''

YiSyllableZyt

\u{a2a8}: ''

YiSyllableZyx

\u{a2a9}: ''

YiSyllableZy

\u{a2aa}: ''

YiSyllableZyp

\u{a2ab}: ''

YiSyllableZyrx

\u{a2ac}: ''

YiSyllableZyr

\u{a2ad}: ''

YiSyllableCit

\u{a2ae}: ''

YiSyllableCix

\u{a2af}: ''

YiSyllableCi

\u{a2b0}: ''

YiSyllableCip

\u{a2b1}: ''

YiSyllableCiet

\u{a2b2}: ''

YiSyllableCiex

\u{a2b3}: ''

YiSyllableCie

\u{a2b4}: ''

YiSyllableCiep

\u{a2b5}: ''

YiSyllableCat

\u{a2b6}: ''

YiSyllableCax

\u{a2b7}: ''

YiSyllableCa

\u{a2b8}: ''

YiSyllableCap

\u{a2b9}: ''

YiSyllableCuox

\u{a2ba}: ''

YiSyllableCuo

\u{a2bb}: ''

YiSyllableCuop

\u{a2bc}: ''

YiSyllableCot

\u{a2bd}: ''

YiSyllableCox

\u{a2be}: ''

YiSyllableCo

\u{a2bf}: ''

YiSyllableCop

\u{a2c0}: ''

YiSyllableCex

\u{a2c1}: ''

YiSyllableCe

\u{a2c2}: ''

YiSyllableCep

\u{a2c3}: ''

YiSyllableCut

\u{a2c4}: ''

YiSyllableCux

\u{a2c5}: ''

YiSyllableCu

\u{a2c6}: ''

YiSyllableCup

\u{a2c7}: ''

YiSyllableCurx

\u{a2c8}: ''

YiSyllableCur

\u{a2c9}: ''

YiSyllableCyt

\u{a2ca}: ''

YiSyllableCyx

\u{a2cb}: ''

YiSyllableCy

\u{a2cc}: ''

YiSyllableCyp

\u{a2cd}: ''

YiSyllableCyrx

\u{a2ce}: ''

YiSyllableCyr

\u{a2cf}: ''

YiSyllableZzit

\u{a2d0}: ''

YiSyllableZzix

\u{a2d1}: ''

YiSyllableZzi

\u{a2d2}: ''

YiSyllableZzip

\u{a2d3}: ''

YiSyllableZziet

\u{a2d4}: ''

YiSyllableZziex

\u{a2d5}: ''

YiSyllableZzie

\u{a2d6}: ''

YiSyllableZziep

\u{a2d7}: ''

YiSyllableZzat

\u{a2d8}: ''

YiSyllableZzax

\u{a2d9}: ''

YiSyllableZza

\u{a2da}: ''

YiSyllableZzap

\u{a2db}: ''

YiSyllableZzox

\u{a2dc}: ''

YiSyllableZzo

\u{a2dd}: ''

YiSyllableZzop

\u{a2de}: ''

YiSyllableZzex

\u{a2df}: ''

YiSyllableZze

\u{a2e0}: ''

YiSyllableZzep

\u{a2e1}: ''

YiSyllableZzux

\u{a2e2}: ''

YiSyllableZzu

\u{a2e3}: ''

YiSyllableZzup

\u{a2e4}: ''

YiSyllableZzurx

\u{a2e5}: ''

YiSyllableZzur

\u{a2e6}: ''

YiSyllableZzyt

\u{a2e7}: ''

YiSyllableZzyx

\u{a2e8}: ''

YiSyllableZzy

\u{a2e9}: ''

YiSyllableZzyp

\u{a2ea}: ''

YiSyllableZzyrx

\u{a2eb}: ''

YiSyllableZzyr

\u{a2ec}: ''

YiSyllableNzit

\u{a2ed}: ''

YiSyllableNzix

\u{a2ee}: ''

YiSyllableNzi

\u{a2ef}: ''

YiSyllableNzip

\u{a2f0}: ''

YiSyllableNziex

\u{a2f1}: ''

YiSyllableNzie

\u{a2f2}: ''

YiSyllableNziep

\u{a2f3}: ''

YiSyllableNzat

\u{a2f4}: ''

YiSyllableNzax

\u{a2f5}: ''

YiSyllableNza

\u{a2f6}: ''

YiSyllableNzap

\u{a2f7}: ''

YiSyllableNzuox

\u{a2f8}: ''

YiSyllableNzuo

\u{a2f9}: ''

YiSyllableNzox

\u{a2fa}: ''

YiSyllableNzop

\u{a2fb}: ''

YiSyllableNzex

\u{a2fc}: ''

YiSyllableNze

\u{a2fd}: ''

YiSyllableNzux

\u{a2fe}: ''

YiSyllableNzu

\u{a2ff}: ''

YiSyllableNzup

\u{a300}: ''

YiSyllableNzurx

\u{a301}: ''

YiSyllableNzur

\u{a302}: ''

YiSyllableNzyt

\u{a303}: ''

YiSyllableNzyx

\u{a304}: ''

YiSyllableNzy

\u{a305}: ''

YiSyllableNzyp

\u{a306}: ''

YiSyllableNzyrx

\u{a307}: ''

YiSyllableNzyr

\u{a308}: ''

YiSyllableSit

\u{a309}: ''

YiSyllableSix

\u{a30a}: ''

YiSyllableSi

\u{a30b}: ''

YiSyllableSip

\u{a30c}: ''

YiSyllableSiex

\u{a30d}: ''

YiSyllableSie

\u{a30e}: ''

YiSyllableSiep

\u{a30f}: ''

YiSyllableSat

\u{a310}: ''

YiSyllableSax

\u{a311}: ''

YiSyllableSa

\u{a312}: ''

YiSyllableSap

\u{a313}: ''

YiSyllableSuox

\u{a314}: ''

YiSyllableSuo

\u{a315}: ''

YiSyllableSuop

\u{a316}: ''

YiSyllableSot

\u{a317}: ''

YiSyllableSox

\u{a318}: ''

YiSyllableSo

\u{a319}: ''

YiSyllableSop

\u{a31a}: ''

YiSyllableSex

\u{a31b}: ''

YiSyllableSe

\u{a31c}: ''

YiSyllableSep

\u{a31d}: ''

YiSyllableSut

\u{a31e}: ''

YiSyllableSux

\u{a31f}: ''

YiSyllableSu

\u{a320}: ''

YiSyllableSup

\u{a321}: ''

YiSyllableSurx

\u{a322}: ''

YiSyllableSur

\u{a323}: ''

YiSyllableSyt

\u{a324}: ''

YiSyllableSyx

\u{a325}: ''

YiSyllableSy

\u{a326}: ''

YiSyllableSyp

\u{a327}: ''

YiSyllableSyrx

\u{a328}: ''

YiSyllableSyr

\u{a329}: ''

YiSyllableSsit

\u{a32a}: ''

YiSyllableSsix

\u{a32b}: ''

YiSyllableSsi

\u{a32c}: ''

YiSyllableSsip

\u{a32d}: ''

YiSyllableSsiex

\u{a32e}: ''

YiSyllableSsie

\u{a32f}: ''

YiSyllableSsiep

\u{a330}: ''

YiSyllableSsat

\u{a331}: ''

YiSyllableSsax

\u{a332}: ''

YiSyllableSsa

\u{a333}: ''

YiSyllableSsap

\u{a334}: ''

YiSyllableSsot

\u{a335}: ''

YiSyllableSsox

\u{a336}: ''

YiSyllableSso

\u{a337}: ''

YiSyllableSsop

\u{a338}: ''

YiSyllableSsex

\u{a339}: ''

YiSyllableSse

\u{a33a}: ''

YiSyllableSsep

\u{a33b}: ''

YiSyllableSsut

\u{a33c}: ''

YiSyllableSsux

\u{a33d}: ''

YiSyllableSsu

\u{a33e}: ''

YiSyllableSsup

\u{a33f}: ''

YiSyllableSsyt

\u{a340}: ''

YiSyllableSsyx

\u{a341}: ''

YiSyllableSsy

\u{a342}: ''

YiSyllableSsyp

\u{a343}: ''

YiSyllableSsyrx

\u{a344}: ''

YiSyllableSsyr

\u{a345}: ''

YiSyllableZhat

\u{a346}: ''

YiSyllableZhax

\u{a347}: ''

YiSyllableZha

\u{a348}: ''

YiSyllableZhap

\u{a349}: ''

YiSyllableZhuox

\u{a34a}: ''

YiSyllableZhuo

\u{a34b}: ''

YiSyllableZhuop

\u{a34c}: ''

YiSyllableZhot

\u{a34d}: ''

YiSyllableZhox

\u{a34e}: ''

YiSyllableZho

\u{a34f}: ''

YiSyllableZhop

\u{a350}: ''

YiSyllableZhet

\u{a351}: ''

YiSyllableZhex

\u{a352}: ''

YiSyllableZhe

\u{a353}: ''

YiSyllableZhep

\u{a354}: ''

YiSyllableZhut

\u{a355}: ''

YiSyllableZhux

\u{a356}: ''

YiSyllableZhu

\u{a357}: ''

YiSyllableZhup

\u{a358}: ''

YiSyllableZhurx

\u{a359}: ''

YiSyllableZhur

\u{a35a}: ''

YiSyllableZhyt

\u{a35b}: ''

YiSyllableZhyx

\u{a35c}: ''

YiSyllableZhy

\u{a35d}: ''

YiSyllableZhyp

\u{a35e}: ''

YiSyllableZhyrx

\u{a35f}: ''

YiSyllableZhyr

\u{a360}: ''

YiSyllableChat

\u{a361}: ''

YiSyllableChax

\u{a362}: ''

YiSyllableCha

\u{a363}: ''

YiSyllableChap

\u{a364}: ''

YiSyllableChuot

\u{a365}: ''

YiSyllableChuox

\u{a366}: ''

YiSyllableChuo

\u{a367}: ''

YiSyllableChuop

\u{a368}: ''

YiSyllableChot

\u{a369}: ''

YiSyllableChox

\u{a36a}: ''

YiSyllableCho

\u{a36b}: ''

YiSyllableChop

\u{a36c}: ''

YiSyllableChet

\u{a36d}: ''

YiSyllableChex

\u{a36e}: ''

YiSyllableChe

\u{a36f}: ''

YiSyllableChep

\u{a370}: ''

YiSyllableChux

\u{a371}: ''

YiSyllableChu

\u{a372}: ''

YiSyllableChup

\u{a373}: ''

YiSyllableChurx

\u{a374}: ''

YiSyllableChur

\u{a375}: ''

YiSyllableChyt

\u{a376}: ''

YiSyllableChyx

\u{a377}: ''

YiSyllableChy

\u{a378}: ''

YiSyllableChyp

\u{a379}: ''

YiSyllableChyrx

\u{a37a}: ''

YiSyllableChyr

\u{a37b}: ''

YiSyllableRrax

\u{a37c}: ''

YiSyllableRra

\u{a37d}: ''

YiSyllableRruox

\u{a37e}: ''

YiSyllableRruo

\u{a37f}: ''

YiSyllableRrot

\u{a380}: ''

YiSyllableRrox

\u{a381}: ''

YiSyllableRro

\u{a382}: ''

YiSyllableRrop

\u{a383}: ''

YiSyllableRret

\u{a384}: ''

YiSyllableRrex

\u{a385}: ''

YiSyllableRre

\u{a386}: ''

YiSyllableRrep

\u{a387}: ''

YiSyllableRrut

\u{a388}: ''

YiSyllableRrux

\u{a389}: ''

YiSyllableRru

\u{a38a}: ''

YiSyllableRrup

\u{a38b}: ''

YiSyllableRrurx

\u{a38c}: ''

YiSyllableRrur

\u{a38d}: ''

YiSyllableRryt

\u{a38e}: ''

YiSyllableRryx

\u{a38f}: ''

YiSyllableRry

\u{a390}: ''

YiSyllableRryp

\u{a391}: ''

YiSyllableRryrx

\u{a392}: ''

YiSyllableRryr

\u{a393}: ''

YiSyllableNrat

\u{a394}: ''

YiSyllableNrax

\u{a395}: ''

YiSyllableNra

\u{a396}: ''

YiSyllableNrap

\u{a397}: ''

YiSyllableNrox

\u{a398}: ''

YiSyllableNro

\u{a399}: ''

YiSyllableNrop

\u{a39a}: ''

YiSyllableNret

\u{a39b}: ''

YiSyllableNrex

\u{a39c}: ''

YiSyllableNre

\u{a39d}: ''

YiSyllableNrep

\u{a39e}: ''

YiSyllableNrut

\u{a39f}: ''

YiSyllableNrux

\u{a3a0}: ''

YiSyllableNru

\u{a3a1}: ''

YiSyllableNrup

\u{a3a2}: ''

YiSyllableNrurx

\u{a3a3}: ''

YiSyllableNrur

\u{a3a4}: ''

YiSyllableNryt

\u{a3a5}: ''

YiSyllableNryx

\u{a3a6}: ''

YiSyllableNry

\u{a3a7}: ''

YiSyllableNryp

\u{a3a8}: ''

YiSyllableNryrx

\u{a3a9}: ''

YiSyllableNryr

\u{a3aa}: ''

YiSyllableShat

\u{a3ab}: ''

YiSyllableShax

\u{a3ac}: ''

YiSyllableSha

\u{a3ad}: ''

YiSyllableShap

\u{a3ae}: ''

YiSyllableShuox

\u{a3af}: ''

YiSyllableShuo

\u{a3b0}: ''

YiSyllableShuop

\u{a3b1}: ''

YiSyllableShot

\u{a3b2}: ''

YiSyllableShox

\u{a3b3}: ''

YiSyllableSho

\u{a3b4}: ''

YiSyllableShop

\u{a3b5}: ''

YiSyllableShet

\u{a3b6}: ''

YiSyllableShex

\u{a3b7}: ''

YiSyllableShe

\u{a3b8}: ''

YiSyllableShep

\u{a3b9}: ''

YiSyllableShut

\u{a3ba}: ''

YiSyllableShux

\u{a3bb}: ''

YiSyllableShu

\u{a3bc}: ''

YiSyllableShup

\u{a3bd}: ''

YiSyllableShurx

\u{a3be}: ''

YiSyllableShur

\u{a3bf}: ''

YiSyllableShyt

\u{a3c0}: ''

YiSyllableShyx

\u{a3c1}: ''

YiSyllableShy

\u{a3c2}: ''

YiSyllableShyp

\u{a3c3}: ''

YiSyllableShyrx

\u{a3c4}: ''

YiSyllableShyr

\u{a3c5}: ''

YiSyllableRat

\u{a3c6}: ''

YiSyllableRax

\u{a3c7}: ''

YiSyllableRa

\u{a3c8}: ''

YiSyllableRap

\u{a3c9}: ''

YiSyllableRuox

\u{a3ca}: ''

YiSyllableRuo

\u{a3cb}: ''

YiSyllableRuop

\u{a3cc}: ''

YiSyllableRot

\u{a3cd}: ''

YiSyllableRox

\u{a3ce}: ''

YiSyllableRo

\u{a3cf}: ''

YiSyllableRop

\u{a3d0}: ''

YiSyllableRex

\u{a3d1}: ''

YiSyllableRe

\u{a3d2}: ''

YiSyllableRep

\u{a3d3}: ''

YiSyllableRut

\u{a3d4}: ''

YiSyllableRux

\u{a3d5}: ''

YiSyllableRu

\u{a3d6}: ''

YiSyllableRup

\u{a3d7}: ''

YiSyllableRurx

\u{a3d8}: ''

YiSyllableRur

\u{a3d9}: ''

YiSyllableRyt

\u{a3da}: ''

YiSyllableRyx

\u{a3db}: ''

YiSyllableRy

\u{a3dc}: ''

YiSyllableRyp

\u{a3dd}: ''

YiSyllableRyrx

\u{a3de}: ''

YiSyllableRyr

\u{a3df}: ''

YiSyllableJit

\u{a3e0}: ''

YiSyllableJix

\u{a3e1}: ''

YiSyllableJi

\u{a3e2}: ''

YiSyllableJip

\u{a3e3}: ''

YiSyllableJiet

\u{a3e4}: ''

YiSyllableJiex

\u{a3e5}: ''

YiSyllableJie

\u{a3e6}: ''

YiSyllableJiep

\u{a3e7}: ''

YiSyllableJuot

\u{a3e8}: ''

YiSyllableJuox

\u{a3e9}: ''

YiSyllableJuo

\u{a3ea}: ''

YiSyllableJuop

\u{a3eb}: ''

YiSyllableJot

\u{a3ec}: ''

YiSyllableJox

\u{a3ed}: ''

YiSyllableJo

\u{a3ee}: ''

YiSyllableJop

\u{a3ef}: ''

YiSyllableJut

\u{a3f0}: ''

YiSyllableJux

\u{a3f1}: ''

YiSyllableJu

\u{a3f2}: ''

YiSyllableJup

\u{a3f3}: ''

YiSyllableJurx

\u{a3f4}: ''

YiSyllableJur

\u{a3f5}: ''

YiSyllableJyt

\u{a3f6}: ''

YiSyllableJyx

\u{a3f7}: ''

YiSyllableJy

\u{a3f8}: ''

YiSyllableJyp

\u{a3f9}: ''

YiSyllableJyrx

\u{a3fa}: ''

YiSyllableJyr

\u{a3fb}: ''

YiSyllableQit

\u{a3fc}: ''

YiSyllableQix

\u{a3fd}: ''

YiSyllableQi

\u{a3fe}: ''

YiSyllableQip

\u{a3ff}: ''

YiSyllableQiet

\u{a400}: ''

YiSyllableQiex

\u{a401}: ''

YiSyllableQie

\u{a402}: ''

YiSyllableQiep

\u{a403}: ''

YiSyllableQuot

\u{a404}: ''

YiSyllableQuox

\u{a405}: ''

YiSyllableQuo

\u{a406}: ''

YiSyllableQuop

\u{a407}: ''

YiSyllableQot

\u{a408}: ''

YiSyllableQox

\u{a409}: ''

YiSyllableQo

\u{a40a}: ''

YiSyllableQop

\u{a40b}: ''

YiSyllableQut

\u{a40c}: ''

YiSyllableQux

\u{a40d}: ''

YiSyllableQu

\u{a40e}: ''

YiSyllableQup

\u{a40f}: ''

YiSyllableQurx

\u{a410}: ''

YiSyllableQur

\u{a411}: ''

YiSyllableQyt

\u{a412}: ''

YiSyllableQyx

\u{a413}: ''

YiSyllableQy

\u{a414}: ''

YiSyllableQyp

\u{a415}: ''

YiSyllableQyrx

\u{a416}: ''

YiSyllableQyr

\u{a417}: ''

YiSyllableJjit

\u{a418}: ''

YiSyllableJjix

\u{a419}: ''

YiSyllableJji

\u{a41a}: ''

YiSyllableJjip

\u{a41b}: ''

YiSyllableJjiet

\u{a41c}: ''

YiSyllableJjiex

\u{a41d}: ''

YiSyllableJjie

\u{a41e}: ''

YiSyllableJjiep

\u{a41f}: ''

YiSyllableJjuox

\u{a420}: ''

YiSyllableJjuo

\u{a421}: ''

YiSyllableJjuop

\u{a422}: ''

YiSyllableJjot

\u{a423}: ''

YiSyllableJjox

\u{a424}: ''

YiSyllableJjo

\u{a425}: ''

YiSyllableJjop

\u{a426}: ''

YiSyllableJjut

\u{a427}: ''

YiSyllableJjux

\u{a428}: ''

YiSyllableJju

\u{a429}: ''

YiSyllableJjup

\u{a42a}: ''

YiSyllableJjurx

\u{a42b}: ''

YiSyllableJjur

\u{a42c}: ''

YiSyllableJjyt

\u{a42d}: ''

YiSyllableJjyx

\u{a42e}: ''

YiSyllableJjy

\u{a42f}: ''

YiSyllableJjyp

\u{a430}: ''

YiSyllableNjit

\u{a431}: ''

YiSyllableNjix

\u{a432}: ''

YiSyllableNji

\u{a433}: ''

YiSyllableNjip

\u{a434}: ''

YiSyllableNjiet

\u{a435}: ''

YiSyllableNjiex

\u{a436}: ''

YiSyllableNjie

\u{a437}: ''

YiSyllableNjiep

\u{a438}: ''

YiSyllableNjuox

\u{a439}: ''

YiSyllableNjuo

\u{a43a}: ''

YiSyllableNjot

\u{a43b}: ''

YiSyllableNjox

\u{a43c}: ''

YiSyllableNjo

\u{a43d}: ''

YiSyllableNjop

\u{a43e}: ''

YiSyllableNjux

\u{a43f}: ''

YiSyllableNju

\u{a440}: ''

YiSyllableNjup

\u{a441}: ''

YiSyllableNjurx

\u{a442}: ''

YiSyllableNjur

\u{a443}: ''

YiSyllableNjyt

\u{a444}: ''

YiSyllableNjyx

\u{a445}: ''

YiSyllableNjy

\u{a446}: ''

YiSyllableNjyp

\u{a447}: ''

YiSyllableNjyrx

\u{a448}: ''

YiSyllableNjyr

\u{a449}: ''

YiSyllableNyit

\u{a44a}: ''

YiSyllableNyix

\u{a44b}: ''

YiSyllableNyi

\u{a44c}: ''

YiSyllableNyip

\u{a44d}: ''

YiSyllableNyiet

\u{a44e}: ''

YiSyllableNyiex

\u{a44f}: ''

YiSyllableNyie

\u{a450}: ''

YiSyllableNyiep

\u{a451}: ''

YiSyllableNyuox

\u{a452}: ''

YiSyllableNyuo

\u{a453}: ''

YiSyllableNyuop

\u{a454}: ''

YiSyllableNyot

\u{a455}: ''

YiSyllableNyox

\u{a456}: ''

YiSyllableNyo

\u{a457}: ''

YiSyllableNyop

\u{a458}: ''

YiSyllableNyut

\u{a459}: ''

YiSyllableNyux

\u{a45a}: ''

YiSyllableNyu

\u{a45b}: ''

YiSyllableNyup

\u{a45c}: ''

YiSyllableXit

\u{a45d}: ''

YiSyllableXix

\u{a45e}: ''

YiSyllableXi

\u{a45f}: ''

YiSyllableXip

\u{a460}: ''

YiSyllableXiet

\u{a461}: ''

YiSyllableXiex

\u{a462}: ''

YiSyllableXie

\u{a463}: ''

YiSyllableXiep

\u{a464}: ''

YiSyllableXuox

\u{a465}: ''

YiSyllableXuo

\u{a466}: ''

YiSyllableXot

\u{a467}: ''

YiSyllableXox

\u{a468}: ''

YiSyllableXo

\u{a469}: ''

YiSyllableXop

\u{a46a}: ''

YiSyllableXyt

\u{a46b}: ''

YiSyllableXyx

\u{a46c}: ''

YiSyllableXy

\u{a46d}: ''

YiSyllableXyp

\u{a46e}: ''

YiSyllableXyrx

\u{a46f}: ''

YiSyllableXyr

\u{a470}: ''

YiSyllableYit

\u{a471}: ''

YiSyllableYix

\u{a472}: ''

YiSyllableYi

\u{a473}: ''

YiSyllableYip

\u{a474}: ''

YiSyllableYiet

\u{a475}: ''

YiSyllableYiex

\u{a476}: ''

YiSyllableYie

\u{a477}: ''

YiSyllableYiep

\u{a478}: ''

YiSyllableYuot

\u{a479}: ''

YiSyllableYuox

\u{a47a}: ''

YiSyllableYuo

\u{a47b}: ''

YiSyllableYuop

\u{a47c}: ''

YiSyllableYot

\u{a47d}: ''

YiSyllableYox

\u{a47e}: ''

YiSyllableYo

\u{a47f}: ''

YiSyllableYop

\u{a480}: ''

YiSyllableYut

\u{a481}: ''

YiSyllableYux

\u{a482}: ''

YiSyllableYu

\u{a483}: ''

YiSyllableYup

\u{a484}: ''

YiSyllableYurx

\u{a485}: ''

YiSyllableYur

\u{a486}: ''

YiSyllableYyt

\u{a487}: ''

YiSyllableYyx

\u{a488}: ''

YiSyllableYy

\u{a489}: ''

YiSyllableYyp

\u{a48a}: ''

YiSyllableYyrx

\u{a48b}: ''

YiSyllableYyr

\u{a48c}: ''

Methods

impl YiSyllables

pub fn new() -> Self

The character with the lowest index in this unicode block

pub fn name(&self) -> String

The character's name, in sentence case